% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_apply_archive_rule}
\alias{accessanalyzer_apply_archive_rule}
\title{Retroactively applies the archive rule to existing findings that meet
the archive rule criteria}
\usage{
accessanalyzer_apply_archive_rule(analyzerArn, clientToken, ruleName)
}
\arguments{
\item{analyzerArn}{[required] The Amazon resource name (ARN) of the analyzer.}

\item{clientToken}{A client token.}

\item{ruleName}{[required] The name of the rule to apply.}
}
\description{
Retroactively applies the archive rule to existing findings that meet
the archive rule criteria.
}
\section{Request syntax}{
\preformatted{svc$apply_archive_rule(
  analyzerArn = "string",
  clientToken = "string",
  ruleName = "string"
)
}
}

\keyword{internal}
