% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_create_analyzer}
\alias{accessanalyzer_create_analyzer}
\title{Creates an analyzer with a zone of trust set to your account}
\usage{
accessanalyzer_create_analyzer(analyzerName, archiveRules, clientToken,
  tags, type)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to create.}

\item{archiveRules}{Specifies the archive rules to add for the analyzer.}

\item{clientToken}{A client token.}

\item{tags}{The tags to apply to the analyzer.}

\item{type}{[required] The zone of trust for the analyzer. You can create only one analyzer per
account per Region.}
}
\description{
Creates an analyzer with a zone of trust set to your account.
}
\section{Request syntax}{
\preformatted{svc$create_analyzer(
  analyzerName = "string",
  archiveRules = list(
    list(
      filter = list(
        list(
          contains = list(
            "string"
          ),
          eq = list(
            "string"
          ),
          exists = TRUE|FALSE,
          neq = list(
            "string"
          )
        )
      ),
      ruleName = "string"
    )
  ),
  clientToken = "string",
  tags = list(
    "string"
  ),
  type = "ACCOUNT"
)
}
}

\keyword{internal}
