% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_create_archive_rule}
\alias{accessanalyzer_create_archive_rule}
\title{Creates an archive rule for the specified analyzer}
\usage{
accessanalyzer_create_archive_rule(analyzerName, clientToken, filter,
  ruleName)
}
\arguments{
\item{analyzerName}{[required] The name of the created analyzer.}

\item{clientToken}{A client token.}

\item{filter}{[required] The criteria for the rule.}

\item{ruleName}{[required] The name of the rule to create.}
}
\description{
Creates an archive rule for the specified analyzer.
}
\section{Request syntax}{
\preformatted{svc$create_archive_rule(
  analyzerName = "string",
  clientToken = "string",
  filter = list(
    list(
      contains = list(
        "string"
      ),
      eq = list(
        "string"
      ),
      exists = TRUE|FALSE,
      neq = list(
        "string"
      )
    )
  ),
  ruleName = "string"
)
}
}

\keyword{internal}
