% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_delete_archive_rule}
\alias{accessanalyzer_delete_archive_rule}
\title{Deletes the specified archive rule}
\usage{
accessanalyzer_delete_archive_rule(analyzerName, clientToken, ruleName)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer that was deleted.}

\item{clientToken}{A client token.}

\item{ruleName}{[required] The name of the rule to delete.}
}
\description{
Deletes the specified archive rule.
}
\section{Request syntax}{
\preformatted{svc$delete_archive_rule(
  analyzerName = "string",
  clientToken = "string",
  ruleName = "string"
)
}
}

\keyword{internal}
