% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_get_analyzed_resource}
\alias{accessanalyzer_get_analyzed_resource}
\title{Retrieves information about an analyzed resource}
\usage{
accessanalyzer_get_analyzed_resource(analyzerArn, resourceArn)
}
\arguments{
\item{analyzerArn}{[required] The ARN of the analyzer to retrieve information from.}

\item{resourceArn}{[required] The ARN of the resource to retrieve information about.}
}
\description{
Retrieves information about an analyzed resource.
}
\section{Request syntax}{
\preformatted{svc$get_analyzed_resource(
  analyzerArn = "string",
  resourceArn = "string"
)
}
}

\keyword{internal}
