% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_list_findings}
\alias{accessanalyzer_list_findings}
\title{Retrieves a list of findings generated by the specified analyzer}
\usage{
accessanalyzer_list_findings(analyzerArn, filter, maxResults, nextToken,
  sort)
}
\arguments{
\item{analyzerArn}{[required] The ARN of the analyzer to retrieve findings from.}

\item{filter}{A filter to match for the findings to return.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token used for pagination of results returned.}

\item{sort}{The sort order for the findings returned.}
}
\description{
Retrieves a list of findings generated by the specified analyzer.
}
\section{Request syntax}{
\preformatted{svc$list_findings(
  analyzerArn = "string",
  filter = list(
    list(
      contains = list(
        "string"
      ),
      eq = list(
        "string"
      ),
      exists = TRUE|FALSE,
      neq = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string",
  sort = list(
    attributeName = "string",
    orderBy = "ASC"|"DESC"
  )
)
}
}

\keyword{internal}
