% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_start_resource_scan}
\alias{accessanalyzer_start_resource_scan}
\title{Starts a scan of the policies applied to the specified resource}
\usage{
accessanalyzer_start_resource_scan(analyzerArn, resourceArn)
}
\arguments{
\item{analyzerArn}{[required] The ARN of the analyzer to use to scan the policies applied to the
specified resource.}

\item{resourceArn}{[required] The ARN of the resource to scan.}
}
\description{
Starts a scan of the policies applied to the specified resource.
}
\section{Request syntax}{
\preformatted{svc$start_resource_scan(
  analyzerArn = "string",
  resourceArn = "string"
)
}
}

\keyword{internal}
