% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_update_archive_rule}
\alias{accessanalyzer_update_archive_rule}
\title{Updates the criteria and values for the specified archive rule}
\usage{
accessanalyzer_update_archive_rule(analyzerName, clientToken, filter,
  ruleName)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to update the archive rules for.}

\item{clientToken}{A client token.}

\item{filter}{[required] A filter to match for the rules to update. Only rules that match the
filter are updated.}

\item{ruleName}{[required] The name of the rule to update.}
}
\description{
Updates the criteria and values for the specified archive rule.
}
\section{Request syntax}{
\preformatted{svc$update_archive_rule(
  analyzerName = "string",
  clientToken = "string",
  filter = list(
    list(
      contains = list(
        "string"
      ),
      eq = list(
        "string"
      ),
      exists = TRUE|FALSE,
      neq = list(
        "string"
      )
    )
  ),
  ruleName = "string"
)
}
}

\keyword{internal}
