% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_update_findings}
\alias{accessanalyzer_update_findings}
\title{Updates findings with the new values provided in the request}
\usage{
accessanalyzer_update_findings(analyzerArn, clientToken, ids,
  resourceArn, status)
}
\arguments{
\item{analyzerArn}{[required] The ARN of the analyzer that generated the findings to update.}

\item{clientToken}{A client token.}

\item{ids}{The IDs of the findings to update.}

\item{resourceArn}{The ARN of the resource identified in the finding.}

\item{status}{[required] The state represents the action to take to update the finding Status.
Use \code{ARCHIVE} to change an Active finding to an Archived finding. Use
\code{ACTIVE} to change an Archived finding to an Active finding.}
}
\description{
Updates findings with the new values provided in the request.
}
\section{Request syntax}{
\preformatted{svc$update_findings(
  analyzerArn = "string",
  clientToken = "string",
  ids = list(
    "string"
  ),
  resourceArn = "string",
  status = "ACTIVE"|"ARCHIVED"
)
}
}

\keyword{internal}
