% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_associate_skill_with_skill_group}
\alias{alexaforbusiness_associate_skill_with_skill_group}
\title{Associates a skill with a skill group}
\usage{
alexaforbusiness_associate_skill_with_skill_group(SkillGroupArn,
  SkillId)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group to associate the skill to. Required.}

\item{SkillId}{[required] The unique identifier of the skill.}
}
\description{
Associates a skill with a skill group.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$associate_skill_with_skill_group(
  SkillGroupArn = "string",
  SkillId = "string"
)
}
}

\keyword{internal}
