% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_associate_skill_with_users}
\alias{alexaforbusiness_associate_skill_with_users}
\title{Makes a private skill available for enrolled users to enable on their
devices}
\usage{
alexaforbusiness_associate_skill_with_users(SkillId)
}
\arguments{
\item{SkillId}{[required] The private skill ID you want to make available to enrolled users.}
}
\description{
Makes a private skill available for enrolled users to enable on their
devices.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$associate_skill_with_users(
  SkillId = "string"
)
}
}

\keyword{internal}
