% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_contact}
\alias{alexaforbusiness_create_contact}
\title{Creates a contact with the specified details}
\usage{
alexaforbusiness_create_contact(DisplayName, FirstName, LastName,
  PhoneNumber, ClientRequestToken)
}
\arguments{
\item{DisplayName}{The name of the contact to display on the console.}

\item{FirstName}{[required] The first name of the contact that is used to call the contact on the
device.}

\item{LastName}{The last name of the contact that is used to call the contact on the
device.}

\item{PhoneNumber}{The phone number of the contact in E.164 format.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures
idempotency.}
}
\description{
Creates a contact with the specified details.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$create_contact(
  DisplayName = "string",
  FirstName = "string",
  LastName = "string",
  PhoneNumber = "string",
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
