% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_profile}
\alias{alexaforbusiness_create_profile}
\title{Creates a new room profile with the specified details}
\usage{
alexaforbusiness_create_profile(ProfileName, Timezone, Address,
  DistanceUnit, TemperatureUnit, WakeWord, Locale, ClientRequestToken,
  SetupModeDisabled, MaxVolumeLimit, PSTNEnabled,
  MeetingRoomConfiguration)
}
\arguments{
\item{ProfileName}{[required] The name of a room profile.}

\item{Timezone}{[required] The time zone used by a room profile.}

\item{Address}{[required] The valid address for the room.}

\item{DistanceUnit}{[required] The distance unit to be used by devices in the profile.}

\item{TemperatureUnit}{[required] The temperature unit to be used by devices in the profile.}

\item{WakeWord}{[required] A wake word for Alexa, Echo, Amazon, or a computer.}

\item{Locale}{The locale of the room profile. (This is currently only available to a
limited preview audience.)}

\item{ClientRequestToken}{The user-specified token that is used during the creation of a profile.}

\item{SetupModeDisabled}{Whether room profile setup is enabled.}

\item{MaxVolumeLimit}{The maximum volume limit for a room profile.}

\item{PSTNEnabled}{Whether PSTN calling is enabled.}

\item{MeetingRoomConfiguration}{The meeting room settings of a room profile.}
}
\description{
Creates a new room profile with the specified details.
}
\section{Request syntax}{
\preformatted{svc$create_profile(
  ProfileName = "string",
  Timezone = "string",
  Address = "string",
  DistanceUnit = "METRIC"|"IMPERIAL",
  TemperatureUnit = "FAHRENHEIT"|"CELSIUS",
  WakeWord = "ALEXA"|"AMAZON"|"ECHO"|"COMPUTER",
  Locale = "string",
  ClientRequestToken = "string",
  SetupModeDisabled = TRUE|FALSE,
  MaxVolumeLimit = 123,
  PSTNEnabled = TRUE|FALSE,
  MeetingRoomConfiguration = list(
    RoomUtilizationMetricsEnabled = TRUE|FALSE,
    EndOfMeetingReminder = list(
      ReminderAtMinutes = list(
        123
      ),
      ReminderType = "ANNOUNCEMENT_TIME_CHECK"|"ANNOUNCEMENT_VARIABLE_TIME_LEFT"|"CHIME"|"KNOCK",
      Enabled = TRUE|FALSE
    ),
    InstantBooking = list(
      DurationInMinutes = 123,
      Enabled = TRUE|FALSE
    ),
    RequireCheckIn = list(
      ReleaseAfterMinutes = 123,
      Enabled = TRUE|FALSE
    )
  )
)
}
}

\keyword{internal}
