% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_skill_group}
\alias{alexaforbusiness_create_skill_group}
\title{Creates a skill group with a specified name and description}
\usage{
alexaforbusiness_create_skill_group(SkillGroupName, Description,
  ClientRequestToken)
}
\arguments{
\item{SkillGroupName}{[required] The name for the skill group.}

\item{Description}{The description for the skill group.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures
idempotency.}
}
\description{
Creates a skill group with a specified name and description.
}
\section{Request syntax}{
\preformatted{svc$create_skill_group(
  SkillGroupName = "string",
  Description = "string",
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
