% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_delete_gateway_group}
\alias{alexaforbusiness_delete_gateway_group}
\title{Deletes a gateway group}
\usage{
alexaforbusiness_delete_gateway_group(GatewayGroupArn)
}
\arguments{
\item{GatewayGroupArn}{[required] The ARN of the gateway group to delete.}
}
\description{
Deletes a gateway group.
}
\section{Request syntax}{
\preformatted{svc$delete_gateway_group(
  GatewayGroupArn = "string"
)
}
}

\keyword{internal}
