% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_delete_room}
\alias{alexaforbusiness_delete_room}
\title{Deletes a room by the room ARN}
\usage{
alexaforbusiness_delete_room(RoomArn)
}
\arguments{
\item{RoomArn}{The ARN of the room to delete. Required.}
}
\description{
Deletes a room by the room ARN.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$delete_room(
  RoomArn = "string"
)
}
}

\keyword{internal}
