% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_conference_provider}
\alias{alexaforbusiness_get_conference_provider}
\title{Gets details about a specific conference provider}
\usage{
alexaforbusiness_get_conference_provider(ConferenceProviderArn)
}
\arguments{
\item{ConferenceProviderArn}{[required] The ARN of the newly created conference provider.}
}
\description{
Gets details about a specific conference provider.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$get_conference_provider(
  ConferenceProviderArn = "string"
)
}
}

\keyword{internal}
