% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_device}
\alias{alexaforbusiness_get_device}
\title{Gets the details of a device by device ARN}
\usage{
alexaforbusiness_get_device(DeviceArn)
}
\arguments{
\item{DeviceArn}{The ARN of the device for which to request details. Required.}
}
\description{
Gets the details of a device by device ARN.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$get_device(
  DeviceArn = "string"
)
}
}

\keyword{internal}
