% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_list_gateways}
\alias{alexaforbusiness_list_gateways}
\title{Retrieves a list of gateway summaries}
\usage{
alexaforbusiness_list_gateways(GatewayGroupArn, NextToken, MaxResults)
}
\arguments{
\item{GatewayGroupArn}{The gateway group ARN for which to list gateways.}

\item{NextToken}{The token used to paginate though multiple pages of gateway summaries.}

\item{MaxResults}{The maximum number of gateway summaries to return. The default is 50.}
}
\description{
Retrieves a list of gateway summaries. Use GetGateway to retrieve
details of a specific gateway. An optional gateway group ARN can be
provided to only retrieve gateway summaries of gateways that are
associated with that gateway group ARN.
}
\section{Request syntax}{
\preformatted{svc$list_gateways(
  GatewayGroupArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
