% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_list_skills}
\alias{alexaforbusiness_list_skills}
\title{Lists all enabled skills in a specific skill group}
\usage{
alexaforbusiness_list_skills(SkillGroupArn, EnablementType, SkillType,
  NextToken, MaxResults)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group for which to list enabled skills.}

\item{EnablementType}{Whether the skill is enabled under the user\'s account.}

\item{SkillType}{Whether the skill is publicly available or is a private skill.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by \code{MaxResults}.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified \code{MaxResults} value, a token is included
in the response so that the remaining results can be retrieved.}
}
\description{
Lists all enabled skills in a specific skill group.
}
\section{Request syntax}{
\preformatted{svc$list_skills(
  SkillGroupArn = "string",
  EnablementType = "ENABLED"|"PENDING",
  SkillType = "PUBLIC"|"PRIVATE"|"ALL",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
