% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_list_skills_store_categories}
\alias{alexaforbusiness_list_skills_store_categories}
\title{Lists all categories in the Alexa skill store}
\usage{
alexaforbusiness_list_skills_store_categories(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The tokens used for pagination.}

\item{MaxResults}{The maximum number of categories returned, per paginated calls.}
}
\description{
Lists all categories in the Alexa skill store.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$list_skills_store_categories(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
