% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_create_app}
\alias{amplify_create_app}
\title{Creates a new Amplify App}
\usage{
amplify_create_app(name, description, repository, platform,
  iamServiceRoleArn, oauthToken, accessToken, environmentVariables,
  enableBranchAutoBuild, enableBasicAuth, basicAuthCredentials,
  customRules, tags, buildSpec, enableAutoBranchCreation,
  autoBranchCreationPatterns, autoBranchCreationConfig)
}
\arguments{
\item{name}{[required] Name for the Amplify App}

\item{description}{Description for an Amplify App}

\item{repository}{Repository for an Amplify App}

\item{platform}{Platform / framework for an Amplify App}

\item{iamServiceRoleArn}{AWS IAM service role for an Amplify App}

\item{oauthToken}{OAuth token for 3rd party source control system for an Amplify App, used
to create webhook and read-only deploy key. OAuth token is not stored.}

\item{accessToken}{Personal Access token for 3rd party source control system for an Amplify
App, used to create webhook and read-only deploy key. Token is not
stored.}

\item{environmentVariables}{Environment variables map for an Amplify App.}

\item{enableBranchAutoBuild}{Enable the auto building of branches for an Amplify App.}

\item{enableBasicAuth}{Enable Basic Authorization for an Amplify App, this will apply to all
branches part of this App.}

\item{basicAuthCredentials}{Credentials for Basic Authorization for an Amplify App.}

\item{customRules}{Custom rewrite / redirect rules for an Amplify App.}

\item{tags}{Tag for an Amplify App}

\item{buildSpec}{BuildSpec for an Amplify App}

\item{enableAutoBranchCreation}{Enables automated branch creation for the Amplify App.}

\item{autoBranchCreationPatterns}{Automated branch creation glob patterns for the Amplify App.}

\item{autoBranchCreationConfig}{Automated branch creation config for the Amplify App.}
}
\description{
Creates a new Amplify App.
}
\section{Request syntax}{
\preformatted{svc$create_app(
  name = "string",
  description = "string",
  repository = "string",
  platform = "WEB",
  iamServiceRoleArn = "string",
  oauthToken = "string",
  accessToken = "string",
  environmentVariables = list(
    "string"
  ),
  enableBranchAutoBuild = TRUE|FALSE,
  enableBasicAuth = TRUE|FALSE,
  basicAuthCredentials = "string",
  customRules = list(
    list(
      source = "string",
      target = "string",
      status = "string",
      condition = "string"
    )
  ),
  tags = list(
    "string"
  ),
  buildSpec = "string",
  enableAutoBranchCreation = TRUE|FALSE,
  autoBranchCreationPatterns = list(
    "string"
  ),
  autoBranchCreationConfig = list(
    stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL",
    framework = "string",
    enableAutoBuild = TRUE|FALSE,
    environmentVariables = list(
      "string"
    ),
    basicAuthCredentials = "string",
    enableBasicAuth = TRUE|FALSE,
    buildSpec = "string"
  )
)
}
}

\keyword{internal}
