% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_create_webhook}
\alias{amplify_create_webhook}
\title{Creates a new webhook on an Amplify app}
\usage{
amplify_create_webhook(appId, branchName, description)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for a branch that is part of an Amplify app.}

\item{description}{The description for a webhook.}
}
\description{
Creates a new webhook on an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$create_webhook(
  appId = "string",
  branchName = "string",
  description = "string"
)
}
}

\keyword{internal}
