% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_generate_access_logs}
\alias{amplify_generate_access_logs}
\title{Retrieve website access logs for a specific time range via a pre-signed
URL}
\usage{
amplify_generate_access_logs(startTime, endTime, domainName, appId)
}
\arguments{
\item{startTime}{The time at which the logs should start, inclusive.}

\item{endTime}{The time at which the logs should end, inclusive.}

\item{domainName}{[required] Name of the domain.}

\item{appId}{[required] Unique Id for an Amplify App.}
}
\description{
Retrieve website access logs for a specific time range via a pre-signed
URL.
}
\section{Request syntax}{
\preformatted{svc$generate_access_logs(
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  domainName = "string",
  appId = "string"
)
}
}

\keyword{internal}
