% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_backend_environment}
\alias{amplify_get_backend_environment}
\title{Retrieves a backend environment for an Amplify App}
\usage{
amplify_get_backend_environment(appId, environmentName)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{environmentName}{[required] Name for the backend environment.}
}
\description{
Retrieves a backend environment for an Amplify App.
}
\section{Request syntax}{
\preformatted{svc$get_backend_environment(
  appId = "string",
  environmentName = "string"
)
}
}

\keyword{internal}
