% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_branch}
\alias{amplify_get_branch}
\title{Returns a branch for an Amplify app}
\usage{
amplify_get_branch(appId, branchName)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for the branch.}
}
\description{
Returns a branch for an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$get_branch(
  appId = "string",
  branchName = "string"
)
}
}

\keyword{internal}
