% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_webhook}
\alias{amplify_get_webhook}
\title{Retrieves webhook info that corresponds to a webhookId}
\usage{
amplify_get_webhook(webhookId)
}
\arguments{
\item{webhookId}{[required] Unique Id for a webhook.}
}
\description{
Retrieves webhook info that corresponds to a webhookId.
}
\section{Request syntax}{
\preformatted{svc$get_webhook(
  webhookId = "string"
)
}
}

\keyword{internal}
