% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_apps}
\alias{amplify_list_apps}
\title{Returns a list of the existing Amplify apps}
\usage{
amplify_list_apps(nextToken, maxResults)
}
\arguments{
\item{nextToken}{A pagination token. If non-null, the pagination token is returned in a
result. Pass its value in another request to retrieve more entries.}

\item{maxResults}{The maximum number of records to list in a single response.}
}
\description{
Returns a list of the existing Amplify apps.
}
\section{Request syntax}{
\preformatted{svc$list_apps(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
