% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_domain_associations}
\alias{amplify_list_domain_associations}
\title{Returns the domain associations for an Amplify app}
\usage{
amplify_list_domain_associations(appId, nextToken, maxResults)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{nextToken}{A pagination token. Set to null to start listing apps from the start. If
non-null, a pagination token is returned in a result. Pass its value in
here to list more projects.}

\item{maxResults}{The maximum number of records to list in a single response.}
}
\description{
Returns the domain associations for an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$list_domain_associations(
  appId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
