% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_tags_for_resource}
\alias{amplify_list_tags_for_resource}
\title{List tags for resource}
\usage{
amplify_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] Resource arn used to list tags.}
}
\description{
List tags for resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
