% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_webhooks}
\alias{amplify_list_webhooks}
\title{List webhooks with an app}
\usage{
amplify_list_webhooks(appId, nextToken, maxResults)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{nextToken}{Pagination token. Set to null to start listing webhooks from start. If
non-null pagination token is returned in a result, then pass its value
in here to list more webhooks.}

\item{maxResults}{Maximum number of records to list in a single response.}
}
\description{
List webhooks with an app.
}
\section{Request syntax}{
\preformatted{svc$list_webhooks(
  appId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
