% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_tag_resource}
\alias{amplify_tag_resource}
\title{Tag resource with tag key and value}
\usage{
amplify_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] Resource arn used to tag resource.}

\item{tags}{[required] Tags used to tag resource.}
}
\description{
Tag resource with tag key and value.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
