% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_untag_resource}
\alias{amplify_untag_resource}
\title{Untag resource with resourceArn}
\usage{
amplify_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] Resource arn used to untag resource.}

\item{tagKeys}{[required] Tag keys used to untag resource.}
}
\description{
Untag resource with resourceArn.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
