% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_update_app}
\alias{amplify_update_app}
\title{Updates an existing Amplify App}
\usage{
amplify_update_app(appId, name, description, platform,
  iamServiceRoleArn, environmentVariables, enableBranchAutoBuild,
  enableBasicAuth, basicAuthCredentials, customRules, buildSpec,
  enableAutoBranchCreation, autoBranchCreationPatterns,
  autoBranchCreationConfig, repository, oauthToken, accessToken)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{name}{Name for an Amplify App.}

\item{description}{Description for an Amplify App.}

\item{platform}{Platform for an Amplify App.}

\item{iamServiceRoleArn}{IAM service role for an Amplify App.}

\item{environmentVariables}{Environment Variables for an Amplify App.}

\item{enableBranchAutoBuild}{Enables branch auto-building for an Amplify App.}

\item{enableBasicAuth}{Enables Basic Authorization for an Amplify App.}

\item{basicAuthCredentials}{Basic Authorization credentials for an Amplify App.}

\item{customRules}{Custom redirect / rewrite rules for an Amplify App.}

\item{buildSpec}{BuildSpec for an Amplify App.}

\item{enableAutoBranchCreation}{Enables automated branch creation for the Amplify App.}

\item{autoBranchCreationPatterns}{Automated branch creation glob patterns for the Amplify App.}

\item{autoBranchCreationConfig}{Automated branch creation branchConfig for the Amplify App.}

\item{repository}{Repository for an Amplify App}

\item{oauthToken}{OAuth token for 3rd party source control system for an Amplify App, used
to create webhook and read-only deploy key. OAuth token is not stored.}

\item{accessToken}{Personal Access token for 3rd party source control system for an Amplify
App, used to create webhook and read-only deploy key. Token is not
stored.}
}
\description{
Updates an existing Amplify App.
}
\section{Request syntax}{
\preformatted{svc$update_app(
  appId = "string",
  name = "string",
  description = "string",
  platform = "WEB",
  iamServiceRoleArn = "string",
  environmentVariables = list(
    "string"
  ),
  enableBranchAutoBuild = TRUE|FALSE,
  enableBasicAuth = TRUE|FALSE,
  basicAuthCredentials = "string",
  customRules = list(
    list(
      source = "string",
      target = "string",
      status = "string",
      condition = "string"
    )
  ),
  buildSpec = "string",
  enableAutoBranchCreation = TRUE|FALSE,
  autoBranchCreationPatterns = list(
    "string"
  ),
  autoBranchCreationConfig = list(
    stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL"|"PULL_REQUEST",
    framework = "string",
    enableAutoBuild = TRUE|FALSE,
    environmentVariables = list(
      "string"
    ),
    basicAuthCredentials = "string",
    enableBasicAuth = TRUE|FALSE,
    buildSpec = "string",
    enablePullRequestPreview = TRUE|FALSE,
    pullRequestEnvironmentName = "string"
  ),
  repository = "string",
  oauthToken = "string",
  accessToken = "string"
)
}
}

\keyword{internal}
