% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_update_webhook}
\alias{amplify_update_webhook}
\title{Update a webhook}
\usage{
amplify_update_webhook(webhookId, branchName, description)
}
\arguments{
\item{webhookId}{[required] Unique Id for a webhook.}

\item{branchName}{Name for a branch, part of an Amplify App.}

\item{description}{Description for a webhook.}
}
\description{
Update a webhook.
}
\section{Request syntax}{
\preformatted{svc$update_webhook(
  webhookId = "string",
  branchName = "string",
  description = "string"
)
}
}

\keyword{internal}
