% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_clone_backend}
\alias{amplifybackend_clone_backend}
\title{This operation clones an existing backend}
\usage{
amplifybackend_clone_backend(AppId, BackendEnvironmentName,
  TargetEnvironmentName)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendEnvironmentName}{[required] The name of the backend environment.}

\item{TargetEnvironmentName}{[required] The name of the destination backend environment to be created.}
}
\description{
This operation clones an existing backend.
}
\section{Request syntax}{
\preformatted{svc$clone_backend(
  AppId = "string",
  BackendEnvironmentName = "string",
  TargetEnvironmentName = "string"
)
}
}

\keyword{internal}
