% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_create_backend}
\alias{amplifybackend_create_backend}
\title{This operation creates a backend for an Amplify app}
\usage{
amplifybackend_create_backend(AppId, AppName, BackendEnvironmentName,
  ResourceConfig, ResourceName)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{AppName}{[required] The name of the app.}

\item{BackendEnvironmentName}{[required] The name of the backend environment.}

\item{ResourceConfig}{The resource configuration for the backend creation request.}

\item{ResourceName}{The name of the resource.}
}
\description{
This operation creates a backend for an Amplify app. Backends are
automatically created at the time of app creation.
}
\section{Request syntax}{
\preformatted{svc$create_backend(
  AppId = "string",
  AppName = "string",
  BackendEnvironmentName = "string",
  ResourceConfig = list(),
  ResourceName = "string"
)
}
}

\keyword{internal}
