% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_create_backend_api}
\alias{amplifybackend_create_backend_api}
\title{Creates a new backend API resource}
\usage{
amplifybackend_create_backend_api(AppId, BackendEnvironmentName,
  ResourceConfig, ResourceName)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendEnvironmentName}{[required] The name of the backend environment.}

\item{ResourceConfig}{[required] The resource configuration for this request.}

\item{ResourceName}{[required] The name of this resource.}
}
\description{
Creates a new backend API resource.
}
\section{Request syntax}{
\preformatted{svc$create_backend_api(
  AppId = "string",
  BackendEnvironmentName = "string",
  ResourceConfig = list(
    AdditionalAuthTypes = list(
      list(
        Mode = "API_KEY"|"AWS_IAM"|"AMAZON_COGNITO_USER_POOLS"|"OPENID_CONNECT",
        Settings = list(
          CognitoUserPoolId = "string",
          Description = "string",
          ExpirationTime = 123.0,
          OpenIDAuthTTL = "string",
          OpenIDClientId = "string",
          OpenIDIatTTL = "string",
          OpenIDIssueURL = "string",
          OpenIDProviderName = "string"
        )
      )
    ),
    ApiName = "string",
    ConflictResolution = list(
      ResolutionStrategy = "OPTIMISTIC_CONCURRENCY"|"LAMBDA"|"AUTOMERGE"|"NONE"
    ),
    DefaultAuthType = list(
      Mode = "API_KEY"|"AWS_IAM"|"AMAZON_COGNITO_USER_POOLS"|"OPENID_CONNECT",
      Settings = list(
        CognitoUserPoolId = "string",
        Description = "string",
        ExpirationTime = 123.0,
        OpenIDAuthTTL = "string",
        OpenIDClientId = "string",
        OpenIDIatTTL = "string",
        OpenIDIssueURL = "string",
        OpenIDProviderName = "string"
      )
    ),
    Service = "string",
    TransformSchema = "string"
  ),
  ResourceName = "string"
)
}
}

\keyword{internal}
