% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_create_backend_auth}
\alias{amplifybackend_create_backend_auth}
\title{Creates a new backend authentication resource}
\usage{
amplifybackend_create_backend_auth(AppId, BackendEnvironmentName,
  ResourceConfig, ResourceName)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendEnvironmentName}{[required] The name of the backend environment.}

\item{ResourceConfig}{[required] The resource configuration for this request object.}

\item{ResourceName}{[required] The name of this resource.}
}
\description{
Creates a new backend authentication resource.
}
\section{Request syntax}{
\preformatted{svc$create_backend_auth(
  AppId = "string",
  BackendEnvironmentName = "string",
  ResourceConfig = list(
    AuthResources = "USER_POOL_ONLY"|"IDENTITY_POOL_AND_USER_POOL",
    IdentityPoolConfigs = list(
      IdentityPoolName = "string",
      UnauthenticatedLogin = TRUE|FALSE
    ),
    Service = "COGNITO",
    UserPoolConfigs = list(
      ForgotPassword = list(
        DeliveryMethod = "EMAIL"|"SMS",
        EmailSettings = list(
          EmailMessage = "string",
          EmailSubject = "string"
        ),
        SmsSettings = list(
          SmsMessage = "string"
        )
      ),
      Mfa = list(
        MFAMode = "ON"|"OFF"|"OPTIONAL",
        Settings = list(
          MfaTypes = list(
            "SMS"|"TOTP"
          ),
          SmsMessage = "string"
        )
      ),
      OAuth = list(
        DomainPrefix = "string",
        OAuthGrantType = "CODE"|"IMPLICIT",
        OAuthScopes = list(
          "PHONE"|"EMAIL"|"OPENID"|"PROFILE"|"AWS_COGNITO_SIGNIN_USER_ADMIN"
        ),
        RedirectSignInURIs = list(
          "string"
        ),
        RedirectSignOutURIs = list(
          "string"
        ),
        SocialProviderSettings = list(
          Facebook = list(
            ClientId = "string",
            ClientSecret = "string"
          ),
          Google = list(
            ClientId = "string",
            ClientSecret = "string"
          ),
          LoginWithAmazon = list(
            ClientId = "string",
            ClientSecret = "string"
          )
        )
      ),
      PasswordPolicy = list(
        AdditionalConstraints = list(
          "REQUIRE_DIGIT"|"REQUIRE_LOWERCASE"|"REQUIRE_SYMBOL"|"REQUIRE_UPPERCASE"
        ),
        MinimumLength = 123.0
      ),
      RequiredSignUpAttributes = list(
        "ADDRESS"|"BIRTHDATE"|"EMAIL"|"FAMILY_NAME"|"GENDER"|"GIVEN_NAME"|"LOCALE"|"MIDDLE_NAME"|"NAME"|"NICKNAME"|"PHONE_NUMBER"|"PICTURE"|"PREFERRED_USERNAME"|"PROFILE"|"UPDATED_AT"|"WEBSITE"|"ZONE_INFO"
      ),
      SignInMethod = "EMAIL"|"EMAIL_AND_PHONE_NUMBER"|"PHONE_NUMBER"|"USERNAME",
      UserPoolName = "string"
    )
  ),
  ResourceName = "string"
)
}
}

\keyword{internal}
