% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_create_backend_config}
\alias{amplifybackend_create_backend_config}
\title{Creates a config object for a backend}
\usage{
amplifybackend_create_backend_config(AppId, BackendManagerAppId)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendManagerAppId}{The app ID for the backend manager.}
}
\description{
Creates a config object for a backend.
}
\section{Request syntax}{
\preformatted{svc$create_backend_config(
  AppId = "string",
  BackendManagerAppId = "string"
)
}
}

\keyword{internal}
