% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_create_token}
\alias{amplifybackend_create_token}
\title{Generates a one-time challenge code to authenticate a user into your
Amplify Admin UI}
\usage{
amplifybackend_create_token(AppId)
}
\arguments{
\item{AppId}{[required] The app ID.}
}
\description{
Generates a one-time challenge code to authenticate a user into your
Amplify Admin UI.
}
\section{Request syntax}{
\preformatted{svc$create_token(
  AppId = "string"
)
}
}

\keyword{internal}
