% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_get_backend}
\alias{amplifybackend_get_backend}
\title{Provides project-level details for your Amplify UI project}
\usage{
amplifybackend_get_backend(AppId, BackendEnvironmentName)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendEnvironmentName}{The name of the backend environment.}
}
\description{
Provides project-level details for your Amplify UI project.
}
\section{Request syntax}{
\preformatted{svc$get_backend(
  AppId = "string",
  BackendEnvironmentName = "string"
)
}
}

\keyword{internal}
