% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_get_backend_api_models}
\alias{amplifybackend_get_backend_api_models}
\title{Generates a model schema for existing backend API resource}
\usage{
amplifybackend_get_backend_api_models(AppId, BackendEnvironmentName,
  ResourceName)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendEnvironmentName}{[required] The name of the backend environment.}

\item{ResourceName}{[required] The name of this resource.}
}
\description{
Generates a model schema for existing backend API resource.
}
\section{Request syntax}{
\preformatted{svc$get_backend_api_models(
  AppId = "string",
  BackendEnvironmentName = "string",
  ResourceName = "string"
)
}
}

\keyword{internal}
