% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_get_backend_auth}
\alias{amplifybackend_get_backend_auth}
\title{Gets backend auth details}
\usage{
amplifybackend_get_backend_auth(AppId, BackendEnvironmentName,
  ResourceName)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendEnvironmentName}{[required] The name of the backend environment.}

\item{ResourceName}{[required] The name of this resource.}
}
\description{
Gets backend auth details.
}
\section{Request syntax}{
\preformatted{svc$get_backend_auth(
  AppId = "string",
  BackendEnvironmentName = "string",
  ResourceName = "string"
)
}
}

\keyword{internal}
