% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_get_backend_job}
\alias{amplifybackend_get_backend_job}
\title{Returns information about a specific job}
\usage{
amplifybackend_get_backend_job(AppId, BackendEnvironmentName, JobId)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendEnvironmentName}{[required] The name of the backend environment.}

\item{JobId}{[required] The ID for the job.}
}
\description{
Returns information about a specific job.
}
\section{Request syntax}{
\preformatted{svc$get_backend_job(
  AppId = "string",
  BackendEnvironmentName = "string",
  JobId = "string"
)
}
}

\keyword{internal}
