% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_get_token}
\alias{amplifybackend_get_token}
\title{Gets the challenge token based on the given appId and sessionId}
\usage{
amplifybackend_get_token(AppId, SessionId)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{SessionId}{[required] The session ID.}
}
\description{
Gets the challenge token based on the given appId and sessionId.
}
\section{Request syntax}{
\preformatted{svc$get_token(
  AppId = "string",
  SessionId = "string"
)
}
}

\keyword{internal}
