% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_list_backend_jobs}
\alias{amplifybackend_list_backend_jobs}
\title{Lists the jobs for the backend of an Amplify app}
\usage{
amplifybackend_list_backend_jobs(AppId, BackendEnvironmentName, JobId,
  MaxResults, NextToken, Operation, Status)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendEnvironmentName}{[required] The name of the backend environment.}

\item{JobId}{The ID for the job.}

\item{MaxResults}{The maximum number of results that you want in the response.}

\item{NextToken}{The token for the next set of results.}

\item{Operation}{Filters the list of response objects to include only those with the
specified operation name.}

\item{Status}{Filters the list of response objects to include only those with the
specified status.}
}
\description{
Lists the jobs for the backend of an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$list_backend_jobs(
  AppId = "string",
  BackendEnvironmentName = "string",
  JobId = "string",
  MaxResults = 123,
  NextToken = "string",
  Operation = "string",
  Status = "string"
)
}
}

\keyword{internal}
