% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_remove_all_backends}
\alias{amplifybackend_remove_all_backends}
\title{Removes all backend environments from your Amplify project}
\usage{
amplifybackend_remove_all_backends(AppId, CleanAmplifyApp)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{CleanAmplifyApp}{Cleans up the Amplify Console app if this value is set to true.}
}
\description{
Removes all backend environments from your Amplify project.
}
\section{Request syntax}{
\preformatted{svc$remove_all_backends(
  AppId = "string",
  CleanAmplifyApp = TRUE|FALSE
)
}
}

\keyword{internal}
