% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_remove_backend_config}
\alias{amplifybackend_remove_backend_config}
\title{Removes the AWS resources that are required to access the Amplify Admin
UI}
\usage{
amplifybackend_remove_backend_config(AppId)
}
\arguments{
\item{AppId}{[required] The app ID.}
}
\description{
Removes the AWS resources that are required to access the Amplify Admin
UI.
}
\section{Request syntax}{
\preformatted{svc$remove_backend_config(
  AppId = "string"
)
}
}

\keyword{internal}
