% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_update_backend_config}
\alias{amplifybackend_update_backend_config}
\title{Updates the AWS resources that are required to access the Amplify Admin
UI}
\usage{
amplifybackend_update_backend_config(AppId, LoginAuthConfig)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{LoginAuthConfig}{Describes the Amazon Cognito configuration for Admin UI access.}
}
\description{
Updates the AWS resources that are required to access the Amplify Admin
UI.
}
\section{Request syntax}{
\preformatted{svc$update_backend_config(
  AppId = "string",
  LoginAuthConfig = list(
    AwsCognitoIdentityPoolId = "string",
    AwsCognitoRegion = "string",
    AwsUserPoolsId = "string",
    AwsUserPoolsWebClientId = "string"
  )
)
}
}

\keyword{internal}
