% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifybackend_operations.R
\name{amplifybackend_update_backend_job}
\alias{amplifybackend_update_backend_job}
\title{Updates a specific job}
\usage{
amplifybackend_update_backend_job(AppId, BackendEnvironmentName, JobId,
  Operation, Status)
}
\arguments{
\item{AppId}{[required] The app ID.}

\item{BackendEnvironmentName}{[required] The name of the backend environment.}

\item{JobId}{[required] The ID for the job.}

\item{Operation}{Filters the list of response objects to include only those with the
specified operation name.}

\item{Status}{Filters the list of response objects to include only those with the
specified status.}
}
\description{
Updates a specific job.
}
\section{Request syntax}{
\preformatted{svc$update_backend_job(
  AppId = "string",
  BackendEnvironmentName = "string",
  JobId = "string",
  Operation = "string",
  Status = "string"
)
}
}

\keyword{internal}
