% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_create_resource}
\alias{apigateway_create_resource}
\title{Creates a Resource resource}
\usage{
apigateway_create_resource(restApiId, parentId, pathPart)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{parentId}{[required] [Required] The parent resource\'s identifier.}

\item{pathPart}{[required] The last path segment for this resource.}
}
\description{
Creates a Resource resource.
}
\section{Request syntax}{
\preformatted{svc$create_resource(
  restApiId = "string",
  parentId = "string",
  pathPart = "string"
)
}
}

\keyword{internal}
