% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_documentation_part}
\alias{apigateway_delete_documentation_part}
\title{Delete documentation part}
\usage{
apigateway_delete_documentation_part(restApiId, documentationPartId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{documentationPartId}{[required] [Required] The identifier of the to-be-deleted documentation part.}
}
\description{
Delete documentation part
}
\section{Request syntax}{
\preformatted{apigateway$delete_documentation_part(
  restApiId = "string",
  documentationPartId = "string"
)
}
}

\keyword{internal}
