% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_method}
\alias{apigateway_delete_method}
\title{Deletes an existing Method resource}
\usage{
apigateway_delete_method(restApiId, resourceId, httpMethod)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] The Resource identifier for the Method resource.}

\item{httpMethod}{[required] [Required] The HTTP verb of the Method resource.}
}
\description{
Deletes an existing Method resource.
}
\section{Request syntax}{
\preformatted{svc$delete_method(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string"
)
}
}

\keyword{internal}
