% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_method_response}
\alias{apigateway_delete_method_response}
\title{Deletes an existing MethodResponse resource}
\usage{
apigateway_delete_method_response(restApiId, resourceId, httpMethod,
  statusCode)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] The Resource identifier for the MethodResponse resource.}

\item{httpMethod}{[required] [Required] The HTTP verb of the Method resource.}

\item{statusCode}{[required] [Required] The status code identifier for the MethodResponse resource.}
}
\description{
Deletes an existing MethodResponse resource.
}
\section{Request syntax}{
\preformatted{apigateway$delete_method_response(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string",
  statusCode = "string"
)
}
}

\keyword{internal}
