% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_flush_stage_authorizers_cache}
\alias{apigateway_flush_stage_authorizers_cache}
\title{Flushes all authorizer cache entries on a stage}
\usage{
apigateway_flush_stage_authorizers_cache(restApiId, stageName)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{stageName}{[required] The name of the stage to flush.}
}
\description{
Flushes all authorizer cache entries on a stage.
}
\section{Request syntax}{
\preformatted{svc$flush_stage_authorizers_cache(
  restApiId = "string",
  stageName = "string"
)
}
}

\keyword{internal}
