% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_api_key}
\alias{apigateway_get_api_key}
\title{Gets information about the current ApiKey resource}
\usage{
apigateway_get_api_key(apiKey, includeValue)
}
\arguments{
\item{apiKey}{[required] [Required] The identifier of the ApiKey resource.}

\item{includeValue}{A boolean flag to specify whether (\code{true}) or not (\code{false}) the result
contains the key value.}
}
\description{
Gets information about the current ApiKey resource.
}
\section{Request syntax}{
\preformatted{apigateway$get_api_key(
  apiKey = "string",
  includeValue = TRUE|FALSE
)
}
}

\keyword{internal}
