% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_documentation_parts}
\alias{apigateway_get_documentation_parts}
\title{Get documentation parts}
\usage{
apigateway_get_documentation_parts(restApiId, type, nameQuery, path,
  position, limit, locationStatus)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{type}{The type of API entities of the to-be-retrieved documentation parts.}

\item{nameQuery}{The name of API entities of the to-be-retrieved documentation parts.}

\item{path}{The path of API entities of the to-be-retrieved documentation parts.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}

\item{locationStatus}{The status of the API documentation parts to retrieve. Valid values are
\code{DOCUMENTED} for retrieving DocumentationPart resources with content and
\code{UNDOCUMENTED} for DocumentationPart resources without content.}
}
\description{

}
\section{Request syntax}{
\preformatted{svc$get_documentation_parts(
  restApiId = "string",
  type = "API"|"AUTHORIZER"|"MODEL"|"RESOURCE"|"METHOD"|"PATH_PARAMETER"|"QUERY_PARAMETER"|"REQUEST_HEADER"|"REQUEST_BODY"|"RESPONSE"|"RESPONSE_HEADER"|"RESPONSE_BODY",
  nameQuery = "string",
  path = "string",
  position = "string",
  limit = 123,
  locationStatus = "DOCUMENTED"|"UNDOCUMENTED"
)
}
}

\keyword{internal}
