% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_domain_names}
\alias{apigateway_get_domain_names}
\title{Represents a collection of DomainName resources}
\usage{
apigateway_get_domain_names(position, limit)
}
\arguments{
\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\description{
Represents a collection of DomainName resources.
}
\section{Request syntax}{
\preformatted{apigateway$get_domain_names(
  position = "string",
  limit = 123
)
}
}

\keyword{internal}
