% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_gateway_response}
\alias{apigateway_get_gateway_response}
\title{Gets a GatewayResponse of a specified response type on the given RestApi}
\usage{
apigateway_get_gateway_response(restApiId, responseType)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{responseType}{[required] [Required]

The response type of the associated GatewayResponse. Valid values are
\itemize{
\item ACCESS\\_DENIED
\item API\\_CONFIGURATION\\_ERROR
\item AUTHORIZER\\_FAILURE
\item AUTHORIZER\\_CONFIGURATION\\_ERROR
\item BAD\\_REQUEST\\_PARAMETERS
\item BAD\\_REQUEST\\_BODY
\item DEFAULT\\_4XX
\item DEFAULT\\_5XX
\item EXPIRED\\_TOKEN
\item INVALID\\_SIGNATURE
\item INTEGRATION\\_FAILURE
\item INTEGRATION\\_TIMEOUT
\item INVALID\\_API\\_KEY
\item MISSING\\_AUTHENTICATION\\_TOKEN
\item QUOTA\\_EXCEEDED
\item REQUEST\\_TOO\\_LARGE
\item RESOURCE\\_NOT\\_FOUND
\item THROTTLED
\item UNAUTHORIZED
\item UNSUPPORTED\\_MEDIA\\_TYPE
}}
}
\description{
Gets a GatewayResponse of a specified response type on the given
RestApi.
}
\section{Request syntax}{
\preformatted{svc$get_gateway_response(
  restApiId = "string",
  responseType = "DEFAULT_4XX"|"DEFAULT_5XX"|"RESOURCE_NOT_FOUND"|"UNAUTHORIZED"|"INVALID_API_KEY"|"ACCESS_DENIED"|"AUTHORIZER_FAILURE"|"AUTHORIZER_CONFIGURATION_ERROR"|"INVALID_SIGNATURE"|"EXPIRED_TOKEN"|"MISSING_AUTHENTICATION_TOKEN"|"INTEGRATION_FAILURE"|"INTEGRATION_TIMEOUT"|"API_CONFIGURATION_ERROR"|"UNSUPPORTED_MEDIA_TYPE"|"BAD_REQUEST_PARAMETERS"|"BAD_REQUEST_BODY"|"REQUEST_TOO_LARGE"|"THROTTLED"|"QUOTA_EXCEEDED"
)
}
}

\keyword{internal}
