% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_model_template}
\alias{apigateway_get_model_template}
\title{Generates a sample mapping template that can be used to transform a
payload into the structure of a model}
\usage{
apigateway_get_model_template(restApiId, modelName)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{modelName}{[required] [Required] The name of the model for which to generate a template.}
}
\description{
Generates a sample mapping template that can be used to transform a
payload into the structure of a model.
}
\section{Request syntax}{
\preformatted{svc$get_model_template(
  restApiId = "string",
  modelName = "string"
)
}
}

\keyword{internal}
