% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_sdk_type}
\alias{apigateway_get_sdk_type}
\title{Get sdk type}
\usage{
apigateway_get_sdk_type(id)
}
\arguments{
\item{id}{[required] [Required] The identifier of the queried SdkType instance.}
}
\description{
Get sdk type
}
\section{Request syntax}{
\preformatted{apigateway$get_sdk_type(
  id = "string"
)
}
}

\keyword{internal}
