% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_untag_resource}
\alias{apigateway_untag_resource}
\title{Removes a tag from a given resource}
\usage{
apigateway_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] [Required] The ARN of a resource that can be tagged. The resource ARN
must be URL-encoded.}

\item{tagKeys}{[required] [Required] The Tag keys to delete.}
}
\description{
Removes a tag from a given resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
