% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_account}
\alias{apigateway_update_account}
\title{Changes information about the current Account resource}
\usage{
apigateway_update_account(patchOperations)
}
\arguments{
\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Changes information about the current Account resource.
}
\section{Request syntax}{
\preformatted{apigateway$update_account(
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
