% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_base_path_mapping}
\alias{apigateway_update_base_path_mapping}
\title{Changes information about the BasePathMapping resource}
\usage{
apigateway_update_base_path_mapping(domainName, basePath,
  patchOperations)
}
\arguments{
\item{domainName}{[required] [Required] The domain name of the BasePathMapping resource to change.}

\item{basePath}{[required] [Required] The base path of the BasePathMapping resource to change.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Changes information about the BasePathMapping resource.
}
\section{Request syntax}{
\preformatted{apigateway$update_base_path_mapping(
  domainName = "string",
  basePath = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
