% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_documentation_part}
\alias{apigateway_update_documentation_part}
\title{Update documentation part}
\usage{
apigateway_update_documentation_part(restApiId, documentationPartId,
  patchOperations)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{documentationPartId}{[required] [Required] The identifier of the to-be-updated documentation part.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Update documentation part
}
\section{Request syntax}{
\preformatted{svc$update_documentation_part(
  restApiId = "string",
  documentationPartId = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
