% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_request_validator}
\alias{apigateway_update_request_validator}
\title{Updates a RequestValidator of a given RestApi}
\usage{
apigateway_update_request_validator(restApiId, requestValidatorId,
  patchOperations)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{requestValidatorId}{[required] [Required] The identifier of RequestValidator to be updated.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Updates a RequestValidator of a given RestApi.
}
\section{Request syntax}{
\preformatted{svc$update_request_validator(
  restApiId = "string",
  requestValidatorId = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
