% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_stage}
\alias{apigateway_update_stage}
\title{Changes information about a Stage resource}
\usage{
apigateway_update_stage(restApiId, stageName, patchOperations)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{stageName}{[required] [Required] The name of the Stage resource to change information about.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Changes information about a Stage resource.
}
\section{Request syntax}{
\preformatted{apigateway$update_stage(
  restApiId = "string",
  stageName = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
