% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_route}
\alias{apigatewayv2_create_route}
\title{Creates a Route for an API}
\usage{
apigatewayv2_create_route(ApiId, ApiKeyRequired, AuthorizationScopes,
  AuthorizationType, AuthorizerId, ModelSelectionExpression,
  OperationName, RequestModels, RequestParameters, RouteKey,
  RouteResponseSelectionExpression, Target)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ApiKeyRequired}{Specifies whether an API key is required for the route.}

\item{AuthorizationScopes}{The authorization scopes supported by this route.}

\item{AuthorizationType}{The authorization type for the route. Valid values are NONE for open
access, AWS\_IAM for using AWS IAM permissions.}

\item{AuthorizerId}{The identifier of the Authorizer resource to be associated with this
route.}

\item{ModelSelectionExpression}{The model selection expression for the route.}

\item{OperationName}{The operation name for the route.}

\item{RequestModels}{The request models for the route.}

\item{RequestParameters}{The request parameters for the route.}

\item{RouteKey}{[required] The route key for the route.}

\item{RouteResponseSelectionExpression}{The route response selection expression for the route.}

\item{Target}{The target for the route.}
}
\description{
Creates a Route for an API.
}
\section{Request syntax}{
\preformatted{apigatewayv2$create_route(
  ApiId = "string",
  ApiKeyRequired = TRUE|FALSE,
  AuthorizationScopes = list(
    "string"
  ),
  AuthorizationType = "NONE"|"AWS_IAM"|"CUSTOM",
  AuthorizerId = "string",
  ModelSelectionExpression = "string",
  OperationName = "string",
  RequestModels = list(
    "string"
  ),
  RequestParameters = list(
    list(
      Required = TRUE|FALSE
    )
  ),
  RouteKey = "string",
  RouteResponseSelectionExpression = "string",
  Target = "string"
)
}
}

\keyword{internal}
