% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_stage}
\alias{apigatewayv2_delete_stage}
\title{Deletes a Stage}
\usage{
apigatewayv2_delete_stage(ApiId, StageName)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{StageName}{[required] The stage name.}
}
\description{
Deletes a Stage.
}
\section{Request syntax}{
\preformatted{apigatewayv2$delete_stage(
  ApiId = "string",
  StageName = "string"
)
}
}

\keyword{internal}
