% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_api_mapping}
\alias{apigatewayv2_get_api_mapping}
\title{The API mapping}
\usage{
apigatewayv2_get_api_mapping(ApiId, ApiMappingId, DomainName)
}
\arguments{
\item{ApiId}{[required] The identifier of the API.}

\item{ApiMappingId}{[required] The API mapping identifier.}

\item{DomainName}{[required] The domain name.}
}
\description{
The API mapping.
}
\section{Request syntax}{
\preformatted{apigatewayv2$get_api_mapping(
  ApiId = "string",
  ApiMappingId = "string",
  DomainName = "string"
)
}
}

\keyword{internal}
