% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_apis}
\alias{apigatewayv2_get_apis}
\title{Gets a collection of Api resources}
\usage{
apigatewayv2_get_apis(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\description{
Gets a collection of Api resources.
}
\section{Request syntax}{
\preformatted{apigatewayv2$get_apis(
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
