% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_authorizer}
\alias{apigatewayv2_get_authorizer}
\title{Gets an Authorizer}
\usage{
apigatewayv2_get_authorizer(ApiId, AuthorizerId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{AuthorizerId}{[required] The authorizer identifier.}
}
\description{
Gets an Authorizer.
}
\section{Request syntax}{
\preformatted{apigatewayv2$get_authorizer(
  ApiId = "string",
  AuthorizerId = "string"
)
}
}

\keyword{internal}
