% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_integration}
\alias{apigatewayv2_get_integration}
\title{Gets an Integration}
\usage{
apigatewayv2_get_integration(ApiId, IntegrationId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}
}
\description{
Gets an Integration.
}
\section{Request syntax}{
\preformatted{svc$get_integration(
  ApiId = "string",
  IntegrationId = "string"
)
}
}

\keyword{internal}
