% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_vpc_link}
\alias{apigatewayv2_get_vpc_link}
\title{Gets a VPC link}
\usage{
apigatewayv2_get_vpc_link(VpcLinkId)
}
\arguments{
\item{VpcLinkId}{[required] The ID of the VPC link.}
}
\description{
Gets a VPC link.
}
\section{Request syntax}{
\preformatted{svc$get_vpc_link(
  VpcLinkId = "string"
)
}
}

\keyword{internal}
