% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_import_api}
\alias{apigatewayv2_import_api}
\title{Imports an API}
\usage{
apigatewayv2_import_api(Basepath, Body, FailOnWarnings)
}
\arguments{
\item{Basepath}{Represents the base path of the imported API. Supported only for HTTP
APIs.}

\item{Body}{[required] The OpenAPI definition. Supported only for HTTP APIs.}

\item{FailOnWarnings}{Specifies whether to rollback the API creation (true) or not (false)
when a warning is encountered. The default value is false.}
}
\description{
Imports an API.
}
\section{Request syntax}{
\preformatted{svc$import_api(
  Basepath = "string",
  Body = "string",
  FailOnWarnings = TRUE|FALSE
)
}
}

\keyword{internal}
