% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_tag_resource}
\alias{apigatewayv2_tag_resource}
\title{Tag an APIGW resource}
\usage{
apigatewayv2_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] AWS resource arn}

\item{Tags}{The key-value map of strings. The valid character set is
[a-zA-Z+-=.\_:/]. The tag key can be up to 128 characters and must not
start with aws:. The tag value can be up to 256 characters..}
}
\description{
Tag an APIGW resource
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
