% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_vpc_link}
\alias{apigatewayv2_update_vpc_link}
\title{Updates a VPC link}
\usage{
apigatewayv2_update_vpc_link(Name, VpcLinkId)
}
\arguments{
\item{Name}{The name of the VPC link.}

\item{VpcLinkId}{[required] The ID of the VPC link.}
}
\description{
Updates a VPC link.
}
\section{Request syntax}{
\preformatted{svc$update_vpc_link(
  Name = "string",
  VpcLinkId = "string"
)
}
}

\keyword{internal}
