% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_create_configuration_profile}
\alias{appconfig_create_configuration_profile}
\title{Information that enables AppConfig to access the configuration source}
\usage{
appconfig_create_configuration_profile(ApplicationId, Name, Description,
  LocationUri, RetrievalRoleArn, Validators, Tags)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{Name}{[required] A name for the configuration profile.}

\item{Description}{A description of the configuration profile.}

\item{LocationUri}{[required] A URI to locate the configuration. You can specify a Systems Manager
(SSM) document, an SSM Parameter Store parameter, or an Amazon S3
object. For an SSM document, specify either the document name in the
format \verb{ssm-document://<Document_name>} or the Amazon Resource Name
(ARN). For a parameter, specify either the parameter name in the format
\verb{ssm-parameter://<Parameter_name>} or the ARN. For an Amazon S3 object,
specify the URI in the following format: \verb{s3://<bucket>/<objectKey> }.
Here is an example: s3://my-bucket/my-app/us-east-1/my-config.json}

\item{RetrievalRoleArn}{The ARN of an IAM role with permission to access the configuration at
the specified LocationUri.}

\item{Validators}{A list of methods for validating the configuration.}

\item{Tags}{Metadata to assign to the configuration profile. Tags help organize and
categorize your AppConfig resources. Each tag consists of a key and an
optional value, both of which you define.}
}
\description{
Information that enables AppConfig to access the configuration source.
Valid configuration sources include Systems Manager (SSM) documents, SSM
Parameter Store parameters, and Amazon S3 objects. A configuration
profile includes the following information.
\itemize{
\item The Uri location of the configuration data.
\item The AWS Identity and Access Management (IAM) role that provides
access to the configuration data.
\item A validator for the configuration data. Available validators include
either a JSON Schema or an AWS Lambda function.
}

For more information, see \href{https://docs.aws.amazon.com/appconfig/latest/userguide/appconfig-creating-configuration-and-profile.html}{Create a Configuration and a Configuration Profile}
in the \emph{AWS AppConfig User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_configuration_profile(
  ApplicationId = "string",
  Name = "string",
  Description = "string",
  LocationUri = "string",
  RetrievalRoleArn = "string",
  Validators = list(
    list(
      Type = "JSON_SCHEMA"|"LAMBDA",
      Content = "string"
    )
  ),
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
