% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_create_hosted_configuration_version}
\alias{appconfig_create_hosted_configuration_version}
\title{Create a new configuration in the AppConfig configuration store}
\usage{
appconfig_create_hosted_configuration_version(ApplicationId,
  ConfigurationProfileId, Description, Content, ContentType,
  LatestVersionNumber)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{ConfigurationProfileId}{[required] The configuration profile ID.}

\item{Description}{A description of the configuration.}

\item{Content}{[required] The content of the configuration or the configuration data.}

\item{ContentType}{[required] A standard MIME type describing the format of the configuration content.
For more information, see Content-Type.}

\item{LatestVersionNumber}{An optional locking token used to prevent race conditions from
overwriting configuration updates when creating a new version. To ensure
your data is not overwritten when creating multiple hosted configuration
versions in rapid succession, specify the version of the latest hosted
configuration version.}
}
\description{
Create a new configuration in the AppConfig configuration store.
}
\section{Request syntax}{
\preformatted{svc$create_hosted_configuration_version(
  ApplicationId = "string",
  ConfigurationProfileId = "string",
  Description = "string",
  Content = raw,
  ContentType = "string",
  LatestVersionNumber = 123
)
}
}

\keyword{internal}
