% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_delete_application}
\alias{appconfig_delete_application}
\title{Delete an application}
\usage{
appconfig_delete_application(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application to delete.}
}
\description{
Delete an application. Deleting an application does not delete a
configuration from a host.
}
\section{Request syntax}{
\preformatted{svc$delete_application(
  ApplicationId = "string"
)
}
}

\keyword{internal}
