% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_delete_configuration_profile}
\alias{appconfig_delete_configuration_profile}
\title{Delete a configuration profile}
\usage{
appconfig_delete_configuration_profile(ApplicationId,
  ConfigurationProfileId)
}
\arguments{
\item{ApplicationId}{[required] The application ID that includes the configuration profile you want to
delete.}

\item{ConfigurationProfileId}{[required] The ID of the configuration profile you want to delete.}
}
\description{
Delete a configuration profile. Deleting a configuration profile does
not delete a configuration from a host.
}
\section{Request syntax}{
\preformatted{svc$delete_configuration_profile(
  ApplicationId = "string",
  ConfigurationProfileId = "string"
)
}
}

\keyword{internal}
