% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_delete_deployment_strategy}
\alias{appconfig_delete_deployment_strategy}
\title{Delete a deployment strategy}
\usage{
appconfig_delete_deployment_strategy(DeploymentStrategyId)
}
\arguments{
\item{DeploymentStrategyId}{[required] The ID of the deployment strategy you want to delete.}
}
\description{
Delete a deployment strategy. Deleting a deployment strategy does not
delete a configuration from a host.
}
\section{Request syntax}{
\preformatted{svc$delete_deployment_strategy(
  DeploymentStrategyId = "string"
)
}
}

\keyword{internal}
