% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_delete_environment}
\alias{appconfig_delete_environment}
\title{Delete an environment}
\usage{
appconfig_delete_environment(ApplicationId, EnvironmentId)
}
\arguments{
\item{ApplicationId}{[required] The application ID that includes the environment you want to delete.}

\item{EnvironmentId}{[required] The ID of the environment you want to delete.}
}
\description{
Delete an environment. Deleting an environment does not delete a
configuration from a host.
}
\section{Request syntax}{
\preformatted{svc$delete_environment(
  ApplicationId = "string",
  EnvironmentId = "string"
)
}
}

\keyword{internal}
