% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_delete_hosted_configuration_version}
\alias{appconfig_delete_hosted_configuration_version}
\title{Delete a version of a configuration from the AppConfig configuration
store}
\usage{
appconfig_delete_hosted_configuration_version(ApplicationId,
  ConfigurationProfileId, VersionNumber)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{ConfigurationProfileId}{[required] The configuration profile ID.}

\item{VersionNumber}{[required] The versions number to delete.}
}
\description{
Delete a version of a configuration from the AppConfig configuration
store.
}
\section{Request syntax}{
\preformatted{svc$delete_hosted_configuration_version(
  ApplicationId = "string",
  ConfigurationProfileId = "string",
  VersionNumber = 123
)
}
}

\keyword{internal}
