% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_get_application}
\alias{appconfig_get_application}
\title{Retrieve information about an application}
\usage{
appconfig_get_application(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application you want to get.}
}
\description{
Retrieve information about an application.
}
\section{Request syntax}{
\preformatted{svc$get_application(
  ApplicationId = "string"
)
}
}

\keyword{internal}
