% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_get_configuration}
\alias{appconfig_get_configuration}
\title{Retrieve information about a configuration}
\usage{
appconfig_get_configuration(Application, Environment, Configuration,
  ClientId, ClientConfigurationVersion)
}
\arguments{
\item{Application}{[required] The application to get.}

\item{Environment}{[required] The environment to get.}

\item{Configuration}{[required] The configuration to get.}

\item{ClientId}{[required] A unique ID to identify the client for the configuration. This ID
enables AppConfig to deploy the configuration in intervals, as defined
in the deployment strategy.}

\item{ClientConfigurationVersion}{The configuration version returned in the most recent GetConfiguration
response.}
}
\description{
Retrieve information about a configuration.
}
\section{Request syntax}{
\preformatted{svc$get_configuration(
  Application = "string",
  Environment = "string",
  Configuration = "string",
  ClientId = "string",
  ClientConfigurationVersion = "string"
)
}
}

\keyword{internal}
