% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_get_configuration_profile}
\alias{appconfig_get_configuration_profile}
\title{Retrieve information about a configuration profile}
\usage{
appconfig_get_configuration_profile(ApplicationId,
  ConfigurationProfileId)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application that includes the configuration profile you
want to get.}

\item{ConfigurationProfileId}{[required] The ID of the configuration profile you want to get.}
}
\description{
Retrieve information about a configuration profile.
}
\section{Request syntax}{
\preformatted{svc$get_configuration_profile(
  ApplicationId = "string",
  ConfigurationProfileId = "string"
)
}
}

\keyword{internal}
