% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_get_deployment}
\alias{appconfig_get_deployment}
\title{Retrieve information about a configuration deployment}
\usage{
appconfig_get_deployment(ApplicationId, EnvironmentId, DeploymentNumber)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application that includes the deployment you want to get.}

\item{EnvironmentId}{[required] The ID of the environment that includes the deployment you want to get.}

\item{DeploymentNumber}{[required] The sequence number of the deployment.}
}
\description{
Retrieve information about a configuration deployment.
}
\section{Request syntax}{
\preformatted{svc$get_deployment(
  ApplicationId = "string",
  EnvironmentId = "string",
  DeploymentNumber = 123
)
}
}

\keyword{internal}
