% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_get_deployment_strategy}
\alias{appconfig_get_deployment_strategy}
\title{Retrieve information about a deployment strategy}
\usage{
appconfig_get_deployment_strategy(DeploymentStrategyId)
}
\arguments{
\item{DeploymentStrategyId}{[required] The ID of the deployment strategy to get.}
}
\description{
Retrieve information about a deployment strategy. A deployment strategy
defines important criteria for rolling out your configuration to the
designated targets. A deployment strategy includes: the overall duration
required, a percentage of targets to receive the deployment during each
interval, an algorithm that defines how percentage grows, and bake time.
}
\section{Request syntax}{
\preformatted{svc$get_deployment_strategy(
  DeploymentStrategyId = "string"
)
}
}

\keyword{internal}
