% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_get_environment}
\alias{appconfig_get_environment}
\title{Retrieve information about an environment}
\usage{
appconfig_get_environment(ApplicationId, EnvironmentId)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application that includes the environment you want to get.}

\item{EnvironmentId}{[required] The ID of the environment you wnat to get.}
}
\description{
Retrieve information about an environment. An environment is a logical
deployment group of AppConfig applications, such as applications in a
\code{Production} environment or in an \code{EU_Region} environment. Each
configuration deployment targets an environment. You can enable one or
more Amazon CloudWatch alarms for an environment. If an alarm is
triggered during a deployment, AppConfig roles back the configuration.
}
\section{Request syntax}{
\preformatted{svc$get_environment(
  ApplicationId = "string",
  EnvironmentId = "string"
)
}
}

\keyword{internal}
