% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_get_hosted_configuration_version}
\alias{appconfig_get_hosted_configuration_version}
\title{Get information about a specific configuration version}
\usage{
appconfig_get_hosted_configuration_version(ApplicationId,
  ConfigurationProfileId, VersionNumber)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{ConfigurationProfileId}{[required] The configuration profile ID.}

\item{VersionNumber}{[required] The version.}
}
\description{
Get information about a specific configuration version.
}
\section{Request syntax}{
\preformatted{svc$get_hosted_configuration_version(
  ApplicationId = "string",
  ConfigurationProfileId = "string",
  VersionNumber = 123
)
}
}

\keyword{internal}
