% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_list_configuration_profiles}
\alias{appconfig_list_configuration_profiles}
\title{Lists the configuration profiles for an application}
\usage{
appconfig_list_configuration_profiles(ApplicationId, MaxResults,
  NextToken)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\description{
Lists the configuration profiles for an application.
}
\section{Request syntax}{
\preformatted{svc$list_configuration_profiles(
  ApplicationId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
