% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_list_deployment_strategies}
\alias{appconfig_list_deployment_strategies}
\title{List deployment strategies}
\usage{
appconfig_list_deployment_strategies(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\description{
List deployment strategies.
}
\section{Request syntax}{
\preformatted{svc$list_deployment_strategies(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
