% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_list_deployments}
\alias{appconfig_list_deployments}
\title{Lists the deployments for an environment}
\usage{
appconfig_list_deployments(ApplicationId, EnvironmentId, MaxResults,
  NextToken)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{EnvironmentId}{[required] The environment ID.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\description{
Lists the deployments for an environment.
}
\section{Request syntax}{
\preformatted{svc$list_deployments(
  ApplicationId = "string",
  EnvironmentId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
