% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_list_hosted_configuration_versions}
\alias{appconfig_list_hosted_configuration_versions}
\title{View a list of configurations stored in the AppConfig configuration
store by version}
\usage{
appconfig_list_hosted_configuration_versions(ApplicationId,
  ConfigurationProfileId, MaxResults, NextToken)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{ConfigurationProfileId}{[required] The configuration profile ID.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\description{
View a list of configurations stored in the AppConfig configuration
store by version.
}
\section{Request syntax}{
\preformatted{svc$list_hosted_configuration_versions(
  ApplicationId = "string",
  ConfigurationProfileId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
