% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_list_tags_for_resource}
\alias{appconfig_list_tags_for_resource}
\title{Retrieves the list of key-value tags assigned to the resource}
\usage{
appconfig_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The resource ARN.}
}
\description{
Retrieves the list of key-value tags assigned to the resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
