% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_start_deployment}
\alias{appconfig_start_deployment}
\title{Starts a deployment}
\usage{
appconfig_start_deployment(ApplicationId, EnvironmentId,
  DeploymentStrategyId, ConfigurationProfileId, ConfigurationVersion,
  Description, Tags)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{EnvironmentId}{[required] The environment ID.}

\item{DeploymentStrategyId}{[required] The deployment strategy ID.}

\item{ConfigurationProfileId}{[required] The configuration profile ID.}

\item{ConfigurationVersion}{[required] The configuration version to deploy.}

\item{Description}{A description of the deployment.}

\item{Tags}{Metadata to assign to the deployment. Tags help organize and categorize
your AppConfig resources. Each tag consists of a key and an optional
value, both of which you define.}
}
\description{
Starts a deployment.
}
\section{Request syntax}{
\preformatted{svc$start_deployment(
  ApplicationId = "string",
  EnvironmentId = "string",
  DeploymentStrategyId = "string",
  ConfigurationProfileId = "string",
  ConfigurationVersion = "string",
  Description = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
