% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_tag_resource}
\alias{appconfig_tag_resource}
\title{Metadata to assign to an AppConfig resource}
\usage{
appconfig_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource for which to retrieve tags.}

\item{Tags}{[required] The key-value string map. The valid character set is [a-zA-Z+-=.\\_:/].
The tag key can be up to 128 characters and must not start with \verb{aws:}.
The tag value can be up to 256 characters.}
}
\description{
Metadata to assign to an AppConfig resource. Tags help organize and
categorize your AppConfig resources. Each tag consists of a key and an
optional value, both of which you define. You can specify a maximum of
50 tags for a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
