% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_untag_resource}
\alias{appconfig_untag_resource}
\title{Deletes a tag key and value from an AppConfig resource}
\usage{
appconfig_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource for which to remove tags.}

\item{TagKeys}{[required] The tag keys to delete.}
}
\description{
Deletes a tag key and value from an AppConfig resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
