% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_update_application}
\alias{appconfig_update_application}
\title{Updates an application}
\usage{
appconfig_update_application(ApplicationId, Name, Description)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{Name}{The name of the application.}

\item{Description}{A description of the application.}
}
\description{
Updates an application.
}
\section{Request syntax}{
\preformatted{svc$update_application(
  ApplicationId = "string",
  Name = "string",
  Description = "string"
)
}
}

\keyword{internal}
