% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_update_configuration_profile}
\alias{appconfig_update_configuration_profile}
\title{Updates a configuration profile}
\usage{
appconfig_update_configuration_profile(ApplicationId,
  ConfigurationProfileId, Name, Description, RetrievalRoleArn, Validators)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{ConfigurationProfileId}{[required] The ID of the configuration profile.}

\item{Name}{The name of the configuration profile.}

\item{Description}{A description of the configuration profile.}

\item{RetrievalRoleArn}{The ARN of an IAM role with permission to access the configuration at
the specified LocationUri.}

\item{Validators}{A list of methods for validating the configuration.}
}
\description{
Updates a configuration profile.
}
\section{Request syntax}{
\preformatted{svc$update_configuration_profile(
  ApplicationId = "string",
  ConfigurationProfileId = "string",
  Name = "string",
  Description = "string",
  RetrievalRoleArn = "string",
  Validators = list(
    list(
      Type = "JSON_SCHEMA"|"LAMBDA",
      Content = "string"
    )
  )
)
}
}

\keyword{internal}
