% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_update_deployment_strategy}
\alias{appconfig_update_deployment_strategy}
\title{Updates a deployment strategy}
\usage{
appconfig_update_deployment_strategy(DeploymentStrategyId, Description,
  DeploymentDurationInMinutes, FinalBakeTimeInMinutes, GrowthFactor,
  GrowthType)
}
\arguments{
\item{DeploymentStrategyId}{[required] The deployment strategy ID.}

\item{Description}{A description of the deployment strategy.}

\item{DeploymentDurationInMinutes}{Total amount of time for a deployment to last.}

\item{FinalBakeTimeInMinutes}{The amount of time AppConfig monitors for alarms before considering the
deployment to be complete and no longer eligible for automatic roll
back.}

\item{GrowthFactor}{The percentage of targets to receive a deployed configuration during
each interval.}

\item{GrowthType}{The algorithm used to define how percentage grows over time.}
}
\description{
Updates a deployment strategy.
}
\section{Request syntax}{
\preformatted{svc$update_deployment_strategy(
  DeploymentStrategyId = "string",
  Description = "string",
  DeploymentDurationInMinutes = 123,
  FinalBakeTimeInMinutes = 123,
  GrowthFactor = 123.0,
  GrowthType = "LINEAR"
)
}
}

\keyword{internal}
