% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_update_environment}
\alias{appconfig_update_environment}
\title{Updates an environment}
\usage{
appconfig_update_environment(ApplicationId, EnvironmentId, Name,
  Description, Monitors)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{EnvironmentId}{[required] The environment ID.}

\item{Name}{The name of the environment.}

\item{Description}{A description of the environment.}

\item{Monitors}{Amazon CloudWatch alarms to monitor during the deployment process.}
}
\description{
Updates an environment.
}
\section{Request syntax}{
\preformatted{svc$update_environment(
  ApplicationId = "string",
  EnvironmentId = "string",
  Name = "string",
  Description = "string",
  Monitors = list(
    list(
      AlarmArn = "string",
      AlarmRoleArn = "string"
    )
  )
)
}
}

\keyword{internal}
