% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_validate_configuration}
\alias{appconfig_validate_configuration}
\title{Uses the validators in a configuration profile to validate a
configuration}
\usage{
appconfig_validate_configuration(ApplicationId, ConfigurationProfileId,
  ConfigurationVersion)
}
\arguments{
\item{ApplicationId}{[required] The application ID.}

\item{ConfigurationProfileId}{[required] The configuration profile ID.}

\item{ConfigurationVersion}{[required] The version of the configuration to validate.}
}
\description{
Uses the validators in a configuration profile to validate a
configuration.
}
\section{Request syntax}{
\preformatted{svc$validate_configuration(
  ApplicationId = "string",
  ConfigurationProfileId = "string",
  ConfigurationVersion = "string"
)
}
}

\keyword{internal}
