% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_create_connector_profile}
\alias{appflow_create_connector_profile}
\title{Creates a new connector profile associated with your AWS account}
\usage{
appflow_create_connector_profile(connectorProfileName, kmsArn,
  connectorType, connectionMode, connectorProfileConfig)
}
\arguments{
\item{connectorProfileName}{[required] The name of the connector profile. The name is unique for each
\code{ConnectorProfile} in your AWS account.}

\item{kmsArn}{The ARN (Amazon Resource Name) of the Key Management Service (KMS) key
you provide for encryption. This is required if you do not want to use
the Amazon AppFlow-managed KMS key. If you don't provide anything here,
Amazon AppFlow uses the Amazon AppFlow-managed KMS key.}

\item{connectorType}{[required] The type of connector, such as Salesforce, Amplitude, and so on.}

\item{connectionMode}{[required] Indicates the connection mode and specifies whether it is public or
private. Private flows use AWS PrivateLink to route data over AWS
infrastructure without exposing it to the public internet.}

\item{connectorProfileConfig}{[required] Defines the connector-specific configuration and credentials.}
}
\description{
Creates a new connector profile associated with your AWS account. There
is a soft quota of 100 connector profiles per AWS account. If you need
more connector profiles than this quota allows, you can submit a request
to the Amazon AppFlow team through the Amazon AppFlow support channel.
}
\section{Request syntax}{
\preformatted{svc$create_connector_profile(
  connectorProfileName = "string",
  kmsArn = "string",
  connectorType = "Salesforce"|"Singular"|"Slack"|"Redshift"|"S3"|"Marketo"|"Googleanalytics"|"Zendesk"|"Servicenow"|"Datadog"|"Trendmicro"|"Snowflake"|"Dynatrace"|"Infornexus"|"Amplitude"|"Veeva"|"EventBridge"|"Upsolver",
  connectionMode = "Public"|"Private",
  connectorProfileConfig = list(
    connectorProfileProperties = list(
      Amplitude = list(),
      Datadog = list(
        instanceUrl = "string"
      ),
      Dynatrace = list(
        instanceUrl = "string"
      ),
      GoogleAnalytics = list(),
      InforNexus = list(
        instanceUrl = "string"
      ),
      Marketo = list(
        instanceUrl = "string"
      ),
      Redshift = list(
        databaseUrl = "string",
        bucketName = "string",
        bucketPrefix = "string",
        roleArn = "string"
      ),
      Salesforce = list(
        instanceUrl = "string",
        isSandboxEnvironment = TRUE|FALSE
      ),
      ServiceNow = list(
        instanceUrl = "string"
      ),
      Singular = list(),
      Slack = list(
        instanceUrl = "string"
      ),
      Snowflake = list(
        warehouse = "string",
        stage = "string",
        bucketName = "string",
        bucketPrefix = "string",
        privateLinkServiceName = "string",
        accountName = "string",
        region = "string"
      ),
      Trendmicro = list(),
      Veeva = list(
        instanceUrl = "string"
      ),
      Zendesk = list(
        instanceUrl = "string"
      )
    ),
    connectorProfileCredentials = list(
      Amplitude = list(
        apiKey = "string",
        secretKey = "string"
      ),
      Datadog = list(
        apiKey = "string",
        applicationKey = "string"
      ),
      Dynatrace = list(
        apiToken = "string"
      ),
      GoogleAnalytics = list(
        clientId = "string",
        clientSecret = "string",
        accessToken = "string",
        refreshToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        )
      ),
      InforNexus = list(
        accessKeyId = "string",
        userId = "string",
        secretAccessKey = "string",
        datakey = "string"
      ),
      Marketo = list(
        clientId = "string",
        clientSecret = "string",
        accessToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        )
      ),
      Redshift = list(
        username = "string",
        password = "string"
      ),
      Salesforce = list(
        accessToken = "string",
        refreshToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        ),
        clientCredentialsArn = "string"
      ),
      ServiceNow = list(
        username = "string",
        password = "string"
      ),
      Singular = list(
        apiKey = "string"
      ),
      Slack = list(
        clientId = "string",
        clientSecret = "string",
        accessToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        )
      ),
      Snowflake = list(
        username = "string",
        password = "string"
      ),
      Trendmicro = list(
        apiSecretKey = "string"
      ),
      Veeva = list(
        username = "string",
        password = "string"
      ),
      Zendesk = list(
        clientId = "string",
        clientSecret = "string",
        accessToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
