% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_delete_connector_profile}
\alias{appflow_delete_connector_profile}
\title{Enables you to delete an existing connector profile}
\usage{
appflow_delete_connector_profile(connectorProfileName, forceDelete)
}
\arguments{
\item{connectorProfileName}{[required] The name of the connector profile. The name is unique for each
\code{ConnectorProfile} in your account.}

\item{forceDelete}{Indicates whether Amazon AppFlow should delete the profile, even if it
is currently in use in one or more flows.}
}
\description{
Enables you to delete an existing connector profile.
}
\section{Request syntax}{
\preformatted{svc$delete_connector_profile(
  connectorProfileName = "string",
  forceDelete = TRUE|FALSE
)
}
}

\keyword{internal}
