% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_delete_flow}
\alias{appflow_delete_flow}
\title{Enables your application to delete an existing flow}
\usage{
appflow_delete_flow(flowName, forceDelete)
}
\arguments{
\item{flowName}{[required] The specified name of the flow. Spaces are not allowed. Use underscores
(\\_) or hyphens (-) only.}

\item{forceDelete}{Indicates whether Amazon AppFlow should delete the flow, even if it is
currently in use.}
}
\description{
Enables your application to delete an existing flow. Before deleting the
flow, Amazon AppFlow validates the request by checking the flow
configuration and status. You can delete flows one at a time.
}
\section{Request syntax}{
\preformatted{svc$delete_flow(
  flowName = "string",
  forceDelete = TRUE|FALSE
)
}
}

\keyword{internal}
