% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_describe_connector_profiles}
\alias{appflow_describe_connector_profiles}
\title{Returns a list of connector-profile details matching the provided
connector-profile names and connector-types}
\usage{
appflow_describe_connector_profiles(connectorProfileNames,
  connectorType, maxResults, nextToken)
}
\arguments{
\item{connectorProfileNames}{The name of the connector profile. The name is unique for each
\code{ConnectorProfile} in the AWS account.}

\item{connectorType}{The type of connector, such as Salesforce, Amplitude, and so on.}

\item{maxResults}{Specifies the maximum number of items that should be returned in the
result set. The default for \code{maxResults} is 20 (for all paginated API
operations).}

\item{nextToken}{The pagination token for the next page of data.}
}
\description{
Returns a list of \code{connector-profile} details matching the provided
\code{connector-profile} names and \code{connector-types}. Both input lists are
optional, and you can use them to filter the result.

If no names or \code{connector-types} are provided, returns all connector
profiles in a paginated form. If there is no match, this operation
returns an empty list.
}
\section{Request syntax}{
\preformatted{svc$describe_connector_profiles(
  connectorProfileNames = list(
    "string"
  ),
  connectorType = "Salesforce"|"Singular"|"Slack"|"Redshift"|"S3"|"Marketo"|"Googleanalytics"|"Zendesk"|"Servicenow"|"Datadog"|"Trendmicro"|"Snowflake"|"Dynatrace"|"Infornexus"|"Amplitude"|"Veeva"|"EventBridge"|"Upsolver",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
