% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_describe_connectors}
\alias{appflow_describe_connectors}
\title{Describes the connectors vended by Amazon AppFlow for specified
connector types}
\usage{
appflow_describe_connectors(connectorTypes, nextToken)
}
\arguments{
\item{connectorTypes}{The type of connector, such as Salesforce, Amplitude, and so on.}

\item{nextToken}{The pagination token for the next page of data.}
}
\description{
Describes the connectors vended by Amazon AppFlow for specified
connector types. If you don't specify a connector type, this operation
describes all connectors vended by Amazon AppFlow. If there are more
connectors than can be returned in one page, the response contains a
\code{nextToken} object, which can be be passed in to the next call to the
\code{\link[=appflow_describe_connectors]{describe_connectors}} API operation to
retrieve the next page.
}
\section{Request syntax}{
\preformatted{svc$describe_connectors(
  connectorTypes = list(
    "Salesforce"|"Singular"|"Slack"|"Redshift"|"S3"|"Marketo"|"Googleanalytics"|"Zendesk"|"Servicenow"|"Datadog"|"Trendmicro"|"Snowflake"|"Dynatrace"|"Infornexus"|"Amplitude"|"Veeva"|"EventBridge"|"Upsolver"
  ),
  nextToken = "string"
)
}
}

\keyword{internal}
