% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_describe_flow}
\alias{appflow_describe_flow}
\title{Provides a description of the specified flow}
\usage{
appflow_describe_flow(flowName)
}
\arguments{
\item{flowName}{[required] The specified name of the flow. Spaces are not allowed. Use underscores
(\\_) or hyphens (-) only.}
}
\description{
Provides a description of the specified flow.
}
\section{Request syntax}{
\preformatted{svc$describe_flow(
  flowName = "string"
)
}
}

\keyword{internal}
