% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_describe_flow_execution_records}
\alias{appflow_describe_flow_execution_records}
\title{Fetches the execution history of the flow}
\usage{
appflow_describe_flow_execution_records(flowName, maxResults, nextToken)
}
\arguments{
\item{flowName}{[required] The specified name of the flow. Spaces are not allowed. Use underscores
(\\_) or hyphens (-) only.}

\item{maxResults}{Specifies the maximum number of items that should be returned in the
result set. The default for \code{maxResults} is 20 (for all paginated API
operations).}

\item{nextToken}{The pagination token for the next page of data.}
}
\description{
Fetches the execution history of the flow.
}
\section{Request syntax}{
\preformatted{svc$describe_flow_execution_records(
  flowName = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
