% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_list_connector_entities}
\alias{appflow_list_connector_entities}
\title{Returns the list of available connector entities supported by Amazon
AppFlow}
\usage{
appflow_list_connector_entities(connectorProfileName, connectorType,
  entitiesPath)
}
\arguments{
\item{connectorProfileName}{The name of the connector profile. The name is unique for each
\code{ConnectorProfile} in the AWS account, and is used to query the
downstream connector.}

\item{connectorType}{The type of connector, such as Salesforce, Amplitude, and so on.}

\item{entitiesPath}{This optional parameter is specific to connector implementation. Some
connectors support multiple levels or categories of entities. You can
find out the list of roots for such providers by sending a request
without the \code{entitiesPath} parameter. If the connector supports entities
at different roots, this initial request returns the list of roots.
Otherwise, this request returns all entities supported by the provider.}
}
\description{
Returns the list of available connector entities supported by Amazon
AppFlow. For example, you can query Salesforce for \emph{Account} and
\emph{Opportunity} entities, or query ServiceNow for the \emph{Incident} entity.
}
\section{Request syntax}{
\preformatted{svc$list_connector_entities(
  connectorProfileName = "string",
  connectorType = "Salesforce"|"Singular"|"Slack"|"Redshift"|"S3"|"Marketo"|"Googleanalytics"|"Zendesk"|"Servicenow"|"Datadog"|"Trendmicro"|"Snowflake"|"Dynatrace"|"Infornexus"|"Amplitude"|"Veeva"|"EventBridge"|"Upsolver",
  entitiesPath = "string"
)
}
}

\keyword{internal}
