% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_list_flows}
\alias{appflow_list_flows}
\title{Lists all of the flows associated with your account}
\usage{
appflow_list_flows(maxResults, nextToken)
}
\arguments{
\item{maxResults}{Specifies the maximum number of items that should be returned in the
result set.}

\item{nextToken}{The pagination token for next page of data.}
}
\description{
Lists all of the flows associated with your account.
}
\section{Request syntax}{
\preformatted{svc$list_flows(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
