% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_list_tags_for_resource}
\alias{appflow_list_tags_for_resource}
\title{Retrieves the tags that are associated with a specified flow}
\usage{
appflow_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the specified flow.}
}
\description{
Retrieves the tags that are associated with a specified flow.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
