% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_start_flow}
\alias{appflow_start_flow}
\title{Activates an existing flow}
\usage{
appflow_start_flow(flowName)
}
\arguments{
\item{flowName}{[required] The specified name of the flow. Spaces are not allowed. Use underscores
(\\_) or hyphens (-) only.}
}
\description{
Activates an existing flow. For on-demand flows, this operation runs the
flow immediately. For schedule and event-triggered flows, this operation
activates the flow.
}
\section{Request syntax}{
\preformatted{svc$start_flow(
  flowName = "string"
)
}
}

\keyword{internal}
