% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_stop_flow}
\alias{appflow_stop_flow}
\title{Deactivates the existing flow}
\usage{
appflow_stop_flow(flowName)
}
\arguments{
\item{flowName}{[required] The specified name of the flow. Spaces are not allowed. Use underscores
(\\_) or hyphens (-) only.}
}
\description{
Deactivates the existing flow. For on-demand flows, this operation
returns an \code{unsupportedOperationException} error message. For schedule
and event-triggered flows, this operation deactivates the flow.
}
\section{Request syntax}{
\preformatted{svc$stop_flow(
  flowName = "string"
)
}
}

\keyword{internal}
