% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_tag_resource}
\alias{appflow_tag_resource}
\title{Applies a tag to the specified flow}
\usage{
appflow_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the flow that you want to tag.}

\item{tags}{[required] The tags used to organize, track, or control access for your flow.}
}
\description{
Applies a tag to the specified flow.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
