% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_untag_resource}
\alias{appflow_untag_resource}
\title{Removes a tag from the specified flow}
\usage{
appflow_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the flow that you want to untag.}

\item{tagKeys}{[required] The tag keys associated with the tag that you want to remove from your
flow.}
}
\description{
Removes a tag from the specified flow.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
