% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appflow_operations.R
\name{appflow_update_connector_profile}
\alias{appflow_update_connector_profile}
\title{Updates a given connector profile associated with your account}
\usage{
appflow_update_connector_profile(connectorProfileName, connectionMode,
  connectorProfileConfig)
}
\arguments{
\item{connectorProfileName}{[required] The name of the connector profile and is unique for each
\code{ConnectorProfile} in the AWS Account.}

\item{connectionMode}{[required] Indicates the connection mode and if it is public or private.}

\item{connectorProfileConfig}{[required] Defines the connector-specific profile configuration and credentials.}
}
\description{
Updates a given connector profile associated with your account.
}
\section{Request syntax}{
\preformatted{svc$update_connector_profile(
  connectorProfileName = "string",
  connectionMode = "Public"|"Private",
  connectorProfileConfig = list(
    connectorProfileProperties = list(
      Amplitude = list(),
      Datadog = list(
        instanceUrl = "string"
      ),
      Dynatrace = list(
        instanceUrl = "string"
      ),
      GoogleAnalytics = list(),
      InforNexus = list(
        instanceUrl = "string"
      ),
      Marketo = list(
        instanceUrl = "string"
      ),
      Redshift = list(
        databaseUrl = "string",
        bucketName = "string",
        bucketPrefix = "string",
        roleArn = "string"
      ),
      Salesforce = list(
        instanceUrl = "string",
        isSandboxEnvironment = TRUE|FALSE
      ),
      ServiceNow = list(
        instanceUrl = "string"
      ),
      Singular = list(),
      Slack = list(
        instanceUrl = "string"
      ),
      Snowflake = list(
        warehouse = "string",
        stage = "string",
        bucketName = "string",
        bucketPrefix = "string",
        privateLinkServiceName = "string",
        accountName = "string",
        region = "string"
      ),
      Trendmicro = list(),
      Veeva = list(
        instanceUrl = "string"
      ),
      Zendesk = list(
        instanceUrl = "string"
      )
    ),
    connectorProfileCredentials = list(
      Amplitude = list(
        apiKey = "string",
        secretKey = "string"
      ),
      Datadog = list(
        apiKey = "string",
        applicationKey = "string"
      ),
      Dynatrace = list(
        apiToken = "string"
      ),
      GoogleAnalytics = list(
        clientId = "string",
        clientSecret = "string",
        accessToken = "string",
        refreshToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        )
      ),
      InforNexus = list(
        accessKeyId = "string",
        userId = "string",
        secretAccessKey = "string",
        datakey = "string"
      ),
      Marketo = list(
        clientId = "string",
        clientSecret = "string",
        accessToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        )
      ),
      Redshift = list(
        username = "string",
        password = "string"
      ),
      Salesforce = list(
        accessToken = "string",
        refreshToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        ),
        clientCredentialsArn = "string"
      ),
      ServiceNow = list(
        username = "string",
        password = "string"
      ),
      Singular = list(
        apiKey = "string"
      ),
      Slack = list(
        clientId = "string",
        clientSecret = "string",
        accessToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        )
      ),
      Snowflake = list(
        username = "string",
        password = "string"
      ),
      Trendmicro = list(
        apiSecretKey = "string"
      ),
      Veeva = list(
        username = "string",
        password = "string"
      ),
      Zendesk = list(
        clientId = "string",
        clientSecret = "string",
        accessToken = "string",
        oAuthRequest = list(
          authCode = "string",
          redirectUri = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
