% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appintegrationsservice_service.R
\name{appintegrationsservice}
\alias{appintegrationsservice}
\title{Amazon AppIntegrations Service}
\usage{
appintegrationsservice(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
The Amazon AppIntegrations APIs are in preview release and are subject
to change.

The Amazon AppIntegrations service enables you to configure and reuse
connections to external applications.

For information about how you can use external applications with Amazon
Connect, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/crm.html}{Set up pre-built integrations}
in the \emph{Amazon Connect Administrator Guide}.
}
\section{Service syntax}{
\preformatted{svc <- appintegrationsservice(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=appintegrationsservice_create_event_integration]{create_event_integration} \tab The Amazon AppIntegrations APIs are in preview release and are subject to change\cr
\link[=appintegrationsservice_delete_event_integration]{delete_event_integration} \tab The Amazon AppIntegrations APIs are in preview release and are subject to change\cr
\link[=appintegrationsservice_get_event_integration]{get_event_integration} \tab The Amazon AppIntegrations APIs are in preview release and are subject to change\cr
\link[=appintegrationsservice_list_event_integration_associations]{list_event_integration_associations} \tab The Amazon AppIntegrations APIs are in preview release and are subject to change\cr
\link[=appintegrationsservice_list_event_integrations]{list_event_integrations} \tab The Amazon AppIntegrations APIs are in preview release and are subject to change\cr
\link[=appintegrationsservice_list_tags_for_resource]{list_tags_for_resource} \tab The Amazon AppIntegrations APIs are in preview release and are subject to change\cr
\link[=appintegrationsservice_tag_resource]{tag_resource} \tab The Amazon AppIntegrations APIs are in preview release and are subject to change\cr
\link[=appintegrationsservice_untag_resource]{untag_resource} \tab The Amazon AppIntegrations APIs are in preview release and are subject to change\cr
\link[=appintegrationsservice_update_event_integration]{update_event_integration} \tab The Amazon AppIntegrations APIs are in preview release and are subject to change
}
}

\examples{
\dontrun{
svc <- appintegrationsservice()
svc$create_event_integration(
  Foo = 123
)
}

}
