% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appintegrationsservice_operations.R
\name{appintegrationsservice_create_event_integration}
\alias{appintegrationsservice_create_event_integration}
\title{The Amazon AppIntegrations APIs are in preview release and are subject
to change}
\usage{
appintegrationsservice_create_event_integration(Name, Description,
  EventFilter, EventBridgeBus, ClientToken, Tags)
}
\arguments{
\item{Name}{[required] The name of the event integration.}

\item{Description}{The description of the event integration.}

\item{EventFilter}{[required] The event filter.}

\item{EventBridgeBus}{[required] The Eventbridge bus.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{Tags}{One or more tags.}
}
\description{
The Amazon AppIntegrations APIs are in preview release and are subject
to change.

Creates an EventIntegration, given a specified name, description, and a
reference to an Amazon Eventbridge bus in your account and a partner
event source that will push events to that bus. No objects are created
in the your account, only metadata that is persisted on the
EventIntegration control plane.
}
\section{Request syntax}{
\preformatted{svc$create_event_integration(
  Name = "string",
  Description = "string",
  EventFilter = list(
    Source = "string"
  ),
  EventBridgeBus = "string",
  ClientToken = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
