% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appintegrationsservice_operations.R
\name{appintegrationsservice_delete_event_integration}
\alias{appintegrationsservice_delete_event_integration}
\title{The Amazon AppIntegrations APIs are in preview release and are subject
to change}
\usage{
appintegrationsservice_delete_event_integration(Name)
}
\arguments{
\item{Name}{[required] The name of the event integration.}
}
\description{
The Amazon AppIntegrations APIs are in preview release and are subject
to change.

Deletes the specified existing event integration. If the event
integration is associated with clients, the request is rejected.
}
\section{Request syntax}{
\preformatted{svc$delete_event_integration(
  Name = "string"
)
}
}

\keyword{internal}
