% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appintegrationsservice_operations.R
\name{appintegrationsservice_get_event_integration}
\alias{appintegrationsservice_get_event_integration}
\title{The Amazon AppIntegrations APIs are in preview release and are subject
to change}
\usage{
appintegrationsservice_get_event_integration(Name)
}
\arguments{
\item{Name}{[required] The name of the event integration.}
}
\description{
The Amazon AppIntegrations APIs are in preview release and are subject
to change.

Return information about the event integration.
}
\section{Request syntax}{
\preformatted{svc$get_event_integration(
  Name = "string"
)
}
}

\keyword{internal}
