% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appintegrationsservice_operations.R
\name{appintegrationsservice_list_event_integration_associations}
\alias{appintegrationsservice_list_event_integration_associations}
\title{The Amazon AppIntegrations APIs are in preview release and are subject
to change}
\usage{
appintegrationsservice_list_event_integration_associations(
  EventIntegrationName, NextToken, MaxResults)
}
\arguments{
\item{EventIntegrationName}{[required] The name of the event integration.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
The Amazon AppIntegrations APIs are in preview release and are subject
to change.

Returns a paginated list of event integration associations in the
account.
}
\section{Request syntax}{
\preformatted{svc$list_event_integration_associations(
  EventIntegrationName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
