% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appintegrationsservice_operations.R
\name{appintegrationsservice_list_tags_for_resource}
\alias{appintegrationsservice_list_tags_for_resource}
\title{The Amazon AppIntegrations APIs are in preview release and are subject
to change}
\usage{
appintegrationsservice_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
The Amazon AppIntegrations APIs are in preview release and are subject
to change.
}
\details{
Lists the tags for the specified resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
