% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appintegrationsservice_operations.R
\name{appintegrationsservice_tag_resource}
\alias{appintegrationsservice_tag_resource}
\title{The Amazon AppIntegrations APIs are in preview release and are subject
to change}
\usage{
appintegrationsservice_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] One or more tags.}
}
\description{
The Amazon AppIntegrations APIs are in preview release and are subject
to change.
}
\details{
Adds the specified tags to the specified resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
