% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appintegrationsservice_operations.R
\name{appintegrationsservice_update_event_integration}
\alias{appintegrationsservice_update_event_integration}
\title{The Amazon AppIntegrations APIs are in preview release and are subject
to change}
\usage{
appintegrationsservice_update_event_integration(Name, Description)
}
\arguments{
\item{Name}{[required] The name of the event integration.}

\item{Description}{The description of the event inegration.}
}
\description{
The Amazon AppIntegrations APIs are in preview release and are subject
to change.

Updates the description of an event integration.
}
\section{Request syntax}{
\preformatted{svc$update_event_integration(
  Name = "string",
  Description = "string"
)
}
}

\keyword{internal}
