% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_create_application}
\alias{applicationinsights_create_application}
\title{Adds an application that is created from a resource group}
\usage{
applicationinsights_create_application(ResourceGroupName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}
}
\description{
Adds an application that is created from a resource group.
}
\section{Request syntax}{
\preformatted{svc$create_application(
  ResourceGroupName = "string"
)
}
}

\keyword{internal}
