% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_component_configuration}
\alias{applicationinsights_describe_component_configuration}
\title{Describes the monitoring configuration of the component}
\usage{
applicationinsights_describe_component_configuration(ResourceGroupName,
  ComponentName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}
}
\description{
Describes the monitoring configuration of the component.
}
\section{Request syntax}{
\preformatted{svc$describe_component_configuration(
  ResourceGroupName = "string",
  ComponentName = "string"
)
}
}

\keyword{internal}
