% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_component_configuration_recommendation}
\alias{applicationinsights_describe_component_configuration_recommendation}
\title{Describes the recommended monitoring configuration of the component}
\usage{
applicationinsights_describe_component_configuration_recommendation(
  ResourceGroupName, ComponentName, Tier)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}

\item{Tier}{[required] The tier of the application component. Supported tiers include
\code{DOT_NET_CORE}, \code{DOT_NET_WORKER}, \code{DOT_NET_WEB}, \code{SQL_SERVER}, and
\code{DEFAULT}.}
}
\description{
Describes the recommended monitoring configuration of the component.
}
\section{Request syntax}{
\preformatted{svc$describe_component_configuration_recommendation(
  ResourceGroupName = "string",
  ComponentName = "string",
  Tier = "DEFAULT"|"DOT_NET_CORE"|"DOT_NET_WORKER"|"DOT_NET_WEB"|"SQL_SERVER"
)
}
}

\keyword{internal}
