% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_problem}
\alias{applicationinsights_describe_problem}
\title{Describes an application problem}
\usage{
applicationinsights_describe_problem(ProblemId)
}
\arguments{
\item{ProblemId}{[required] The ID of the problem.}
}
\description{
Describes an application problem.
}
\section{Request syntax}{
\preformatted{svc$describe_problem(
  ProblemId = "string"
)
}
}

\keyword{internal}
