% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_delete_virtual_node}
\alias{appmesh_delete_virtual_node}
\title{Deletes an existing virtual node}
\usage{
appmesh_delete_virtual_node(meshName, virtualNodeName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh in which to delete the virtual node.}

\item{virtualNodeName}{[required] The name of the virtual node to delete.}
}
\description{
Deletes an existing virtual node.
}
\section{Request syntax}{
\preformatted{appmesh$delete_virtual_node(
  meshName = "string",
  virtualNodeName = "string"
)
}
}

\keyword{internal}
