% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_delete_virtual_router}
\alias{appmesh_delete_virtual_router}
\title{Deletes an existing virtual router}
\usage{
appmesh_delete_virtual_router(meshName, virtualRouterName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to delete the virtual router in.}

\item{virtualRouterName}{[required] The name of the virtual router to delete.}
}
\description{
Deletes an existing virtual router.
}
\details{
You must delete any routes associated with the virtual router before you
can delete the router itself.
}
\section{Request syntax}{
\preformatted{svc$delete_virtual_router(
  meshName = "string",
  virtualRouterName = "string"
)
}
}

\keyword{internal}
