% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_delete_virtual_service}
\alias{appmesh_delete_virtual_service}
\title{Deletes an existing virtual service}
\usage{
appmesh_delete_virtual_service(meshName, virtualServiceName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to delete the virtual service in.}

\item{virtualServiceName}{[required] The name of the virtual service to delete.}
}
\description{
Deletes an existing virtual service.
}
\section{Request syntax}{
\preformatted{appmesh$delete_virtual_service(
  meshName = "string",
  virtualServiceName = "string"
)
}
}

\keyword{internal}
