% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_describe_mesh}
\alias{appmesh_describe_mesh}
\title{Describes an existing service mesh}
\usage{
appmesh_describe_mesh(meshName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to describe.}
}
\description{
Describes an existing service mesh.
}
\section{Request syntax}{
\preformatted{appmesh$describe_mesh(
  meshName = "string"
)
}
}

\keyword{internal}
