% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_describe_virtual_node}
\alias{appmesh_describe_virtual_node}
\title{Describes an existing virtual node}
\usage{
appmesh_describe_virtual_node(meshName, virtualNodeName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh that the virtual node resides in.}

\item{virtualNodeName}{[required] The name of the virtual node to describe.}
}
\description{
Describes an existing virtual node.
}
\section{Request syntax}{
\preformatted{svc$describe_virtual_node(
  meshName = "string",
  virtualNodeName = "string"
)
}
}

\keyword{internal}
