% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_describe_virtual_router}
\alias{appmesh_describe_virtual_router}
\title{Describes an existing virtual router}
\usage{
appmesh_describe_virtual_router(meshName, virtualRouterName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh in which the virtual router resides.}

\item{virtualRouterName}{[required] The name of the virtual router to describe.}
}
\description{
Describes an existing virtual router.
}
\section{Request syntax}{
\preformatted{appmesh$describe_virtual_router(
  meshName = "string",
  virtualRouterName = "string"
)
}
}

\keyword{internal}
