% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_describe_virtual_service}
\alias{appmesh_describe_virtual_service}
\title{Describes an existing virtual service}
\usage{
appmesh_describe_virtual_service(meshName, virtualServiceName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh that the virtual service resides in.}

\item{virtualServiceName}{[required] The name of the virtual service to describe.}
}
\description{
Describes an existing virtual service.
}
\section{Request syntax}{
\preformatted{appmesh$describe_virtual_service(
  meshName = "string",
  virtualServiceName = "string"
)
}
}

\keyword{internal}
