% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_update_mesh}
\alias{appmesh_update_mesh}
\title{Updates an existing service mesh}
\usage{
appmesh_update_mesh(clientToken, meshName, spec)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Up to 36 letters, numbers, hyphens, and
underscores are allowed.}

\item{meshName}{[required] The name of the service mesh to update.}

\item{spec}{The service mesh specification to apply.}
}
\description{
Updates an existing service mesh.
}
\section{Request syntax}{
\preformatted{appmesh$update_mesh(
  clientToken = "string",
  meshName = "string",
  spec = list(
    egressFilter = list(
      type = "ALLOW_ALL"|"DROP_ALL"
    )
  )
)
}
}

\keyword{internal}
