% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_update_virtual_router}
\alias{appmesh_update_virtual_router}
\title{Updates an existing virtual router in a specified service mesh}
\usage{
appmesh_update_virtual_router(clientToken, meshName, spec,
  virtualRouterName)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Up to 36 letters, numbers, hyphens, and
underscores are allowed.}

\item{meshName}{[required] The name of the service mesh that the virtual router resides in.}

\item{spec}{[required] The new virtual router specification to apply. This overwrites the
existing data.}

\item{virtualRouterName}{[required] The name of the virtual router to update.}
}
\description{
Updates an existing virtual router in a specified service mesh.
}
\section{Request syntax}{
\preformatted{appmesh$update_virtual_router(
  clientToken = "string",
  meshName = "string",
  spec = list(
    listeners = list(
      list(
        portMapping = list(
          port = 123,
          protocol = "http"|"tcp"
        )
      )
    )
  ),
  virtualRouterName = "string"
)
}
}

\keyword{internal}
