% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_associate_resource}
\alias{appregistry_associate_resource}
\title{Associates a resource with an application}
\usage{
appregistry_associate_resource(application, resourceType, resource)
}
\arguments{
\item{application}{[required] The name or ID of the application.}

\item{resourceType}{[required] The type of resource of which the application will be associated.}

\item{resource}{[required] The name or ID of the resource of which the application will be
associated.}
}
\description{
Associates a resource with an application. Both the resource and the
application can be specified either by ID or name.
}
\section{Request syntax}{
\preformatted{svc$associate_resource(
  application = "string",
  resourceType = "CFN_STACK",
  resource = "string"
)
}
}

\keyword{internal}
