% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_delete_attribute_group}
\alias{appregistry_delete_attribute_group}
\title{Deletes an attribute group, specified either by its attribute group ID
or name}
\usage{
appregistry_delete_attribute_group(attributeGroup)
}
\arguments{
\item{attributeGroup}{[required] The name or ID of the attribute group that holds the attributes to
describe the application.}
}
\description{
Deletes an attribute group, specified either by its attribute group ID
or name.
}
\section{Request syntax}{
\preformatted{svc$delete_attribute_group(
  attributeGroup = "string"
)
}
}

\keyword{internal}
