% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_disassociate_attribute_group}
\alias{appregistry_disassociate_attribute_group}
\title{Disassociates an attribute group from an application to remove the extra
attributes contained in the attribute group from the application's
metadata}
\usage{
appregistry_disassociate_attribute_group(application, attributeGroup)
}
\arguments{
\item{application}{[required] The name or ID of the application.}

\item{attributeGroup}{[required] The name or ID of the attribute group that holds the attributes to
describe the application.}
}
\description{
Disassociates an attribute group from an application to remove the extra
attributes contained in the attribute group from the application's
metadata. This operation reverts \code{AssociateAttributeGroup}.
}
\section{Request syntax}{
\preformatted{svc$disassociate_attribute_group(
  application = "string",
  attributeGroup = "string"
)
}
}

\keyword{internal}
