% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_list_associated_attribute_groups}
\alias{appregistry_list_associated_attribute_groups}
\title{Lists all attribute groups that are associated with specified
application}
\usage{
appregistry_list_associated_attribute_groups(application, nextToken,
  maxResults)
}
\arguments{
\item{application}{[required] The name or ID of the application.}

\item{nextToken}{The token to use to get the next page of results after a previous API
call.}

\item{maxResults}{The upper bound of the number of results to return (cannot exceed 25).
If this parameter is omitted, it defaults to 25. This value is optional.}
}
\description{
Lists all attribute groups that are associated with specified
application. Results are paginated.
}
\section{Request syntax}{
\preformatted{svc$list_associated_attribute_groups(
  application = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
