% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_list_attribute_groups}
\alias{appregistry_list_attribute_groups}
\title{Lists all attribute groups which you have access to}
\usage{
appregistry_list_attribute_groups(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token to use to get the next page of results after a previous API
call.}

\item{maxResults}{The upper bound of the number of results to return (cannot exceed 25).
If this parameter is omitted, it defaults to 25. This value is optional.}
}
\description{
Lists all attribute groups which you have access to. Results are
paginated.
}
\section{Request syntax}{
\preformatted{svc$list_attribute_groups(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
