% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_list_tags_for_resource}
\alias{appregistry_list_tags_for_resource}
\title{Lists all of the tags on the resource}
\usage{
appregistry_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon resource name (ARN) that specifies the resource.}
}
\description{
Lists all of the tags on the resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
