% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_sync_resource}
\alias{appregistry_sync_resource}
\title{Syncs the resource with what is currently recorded in App registry}
\usage{
appregistry_sync_resource(resourceType, resource)
}
\arguments{
\item{resourceType}{[required] The type of resource of which the application will be associated.}

\item{resource}{[required] An entity you can work with and specify with a name or ID. Examples
include an Amazon EC2 instance, an AWS CloudFormation stack, or an
Amazon S3 bucket.}
}
\description{
Syncs the resource with what is currently recorded in App registry.
Specifically, the resource’s App registry system tags are synced with
its associated application. The resource is removed if it is not
associated with the application. The caller must have permissions to
read and update the resource.
}
\section{Request syntax}{
\preformatted{svc$sync_resource(
  resourceType = "CFN_STACK",
  resource = "string"
)
}
}

\keyword{internal}
