% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_tag_resource}
\alias{appregistry_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified resource}
\usage{
appregistry_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon resource name (ARN) that specifies the resource.}

\item{tags}{[required] The new or modified tags for the resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified resource.
}
\details{
Each tag consists of a key and an optional value. If a tag with the same
key is already associated with the resource, this action updates its
value.

This operation returns an empty response if the call was successful.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
