% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_untag_resource}
\alias{appregistry_untag_resource}
\title{Removes tags from a resource}
\usage{
appregistry_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon resource name (ARN) that specifies the resource.}

\item{tagKeys}{[required] A list of the tag keys to remove from the specified resource.}
}
\description{
Removes tags from a resource.
}
\details{
This operation returns an empty response if the call was successful.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
