% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_update_application}
\alias{appregistry_update_application}
\title{Updates an existing application with new attributes}
\usage{
appregistry_update_application(application, name, description)
}
\arguments{
\item{application}{[required] The name or ID of the application that will be updated.}

\item{name}{The new name of the application. The name must be unique in the region
in which you are updating the application.}

\item{description}{The new description of the application.}
}
\description{
Updates an existing application with new attributes.
}
\section{Request syntax}{
\preformatted{svc$update_application(
  application = "string",
  name = "string",
  description = "string"
)
}
}

\keyword{internal}
