% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_create_usage_report_subscription}
\alias{appstream_create_usage_report_subscription}
\title{Creates a usage report subscription}
\usage{
appstream_create_usage_report_subscription()
}
\description{
Creates a usage report subscription. Usage reports are generated daily.
}
\section{Request syntax}{
\preformatted{svc$create_usage_report_subscription()
}
}

\keyword{internal}
