% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_disassociate_fleet}
\alias{appstream_disassociate_fleet}
\title{Disassociates the specified fleet from the specified stack}
\usage{
appstream_disassociate_fleet(FleetName, StackName)
}
\arguments{
\item{FleetName}{[required] The name of the fleet.}

\item{StackName}{[required] The name of the stack.}
}
\description{
Disassociates the specified fleet from the specified stack.
}
\section{Request syntax}{
\preformatted{appstream$disassociate_fleet(
  FleetName = "string",
  StackName = "string"
)
}
}

\keyword{internal}
