% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_get_api_cache}
\alias{appsync_get_api_cache}
\title{Retrieves an ApiCache object}
\usage{
appsync_get_api_cache(apiId)
}
\arguments{
\item{apiId}{[required] The API ID.}
}
\description{
Retrieves an \code{ApiCache} object.
}
\section{Request syntax}{
\preformatted{svc$get_api_cache(
  apiId = "string"
)
}
}

\keyword{internal}
