% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_update_type}
\alias{appsync_update_type}
\title{Updates a Type object}
\usage{
appsync_update_type(apiId, typeName, definition, format)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The new type name.}

\item{definition}{The new definition.}

\item{format}{[required] The new type format: SDL or JSON.}
}
\description{
Updates a \code{Type} object.
}
\section{Request syntax}{
\preformatted{appsync$update_type(
  apiId = "string",
  typeName = "string",
  definition = "string",
  format = "SDL"|"JSON"
)
}
}

\keyword{internal}
