% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_batch_get_named_query}
\alias{athena_batch_get_named_query}
\title{Returns the details of a single named query or a list of up to 50
queries, which you provide as an array of query ID strings}
\usage{
athena_batch_get_named_query(NamedQueryIds)
}
\arguments{
\item{NamedQueryIds}{[required] An array of query IDs.}
}
\description{
Returns the details of a single named query or a list of up to 50
queries, which you provide as an array of query ID strings. Use
ListNamedQueries to get the list of named query IDs. If information
could not be retrieved for a submitted query ID, information about the
query ID submitted is listed under UnprocessedNamedQueryId. Named
queries are different from executed queries. Use BatchGetQueryExecution
to get details about each unique query execution, and
ListQueryExecutions to get a list of query execution IDs.
}
\section{Request syntax}{
\preformatted{athena$batch_get_named_query(
  NamedQueryIds = list(
    "string"
  )
)
}
}

\keyword{internal}
