% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_delete_named_query}
\alias{athena_delete_named_query}
\title{Deletes the named query if you have access to the workgroup in which the
query was saved}
\usage{
athena_delete_named_query(NamedQueryId)
}
\arguments{
\item{NamedQueryId}{[required] The unique ID of the query to delete.}
}
\description{
Deletes the named query if you have access to the workgroup in which the
query was saved.
}
\details{
For code samples using the AWS SDK for Java, see \href{https://docs.aws.amazon.com/athena/latest/ug/code-samples.html}{Examples and Code Samples}
in the \emph{Amazon Athena User Guide}.
}
\section{Request syntax}{
\preformatted{svc$delete_named_query(
  NamedQueryId = "string"
)
}
}

\keyword{internal}
