% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_delete_work_group}
\alias{athena_delete_work_group}
\title{Deletes the workgroup with the specified name}
\usage{
athena_delete_work_group(WorkGroup, RecursiveDeleteOption)
}
\arguments{
\item{WorkGroup}{[required] The unique name of the workgroup to delete.}

\item{RecursiveDeleteOption}{The option to delete the workgroup and its contents even if the
workgroup contains any named queries.}
}
\description{
Deletes the workgroup with the specified name. The primary workgroup
cannot be deleted.
}
\section{Request syntax}{
\preformatted{athena$delete_work_group(
  WorkGroup = "string",
  RecursiveDeleteOption = TRUE|FALSE
)
}
}

\keyword{internal}
