% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_data_catalog}
\alias{athena_get_data_catalog}
\title{Returns the specified data catalog}
\usage{
athena_get_data_catalog(Name)
}
\arguments{
\item{Name}{[required] The name of the data catalog to return.}
}
\description{
Returns the specified data catalog.
}
\section{Request syntax}{
\preformatted{svc$get_data_catalog(
  Name = "string"
)
}
}

\keyword{internal}
