% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_named_query}
\alias{athena_get_named_query}
\title{Returns information about a single query}
\usage{
athena_get_named_query(NamedQueryId)
}
\arguments{
\item{NamedQueryId}{[required] The unique ID of the query. Use ListNamedQueries to get query IDs.}
}
\description{
Returns information about a single query.
}
\section{Request syntax}{
\preformatted{athena$get_named_query(
  NamedQueryId = "string"
)
}
}

\keyword{internal}
