% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_work_groups}
\alias{athena_list_work_groups}
\title{Lists available workgroups for the account}
\usage{
athena_list_work_groups(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{A token to be used by the next request if this request is truncated.}

\item{MaxResults}{The maximum number of workgroups to return in this request.}
}
\description{
Lists available workgroups for the account.
}
\section{Request syntax}{
\preformatted{athena$list_work_groups(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
