% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_stop_query_execution}
\alias{athena_stop_query_execution}
\title{Stops a query execution}
\usage{
athena_stop_query_execution(QueryExecutionId)
}
\arguments{
\item{QueryExecutionId}{[required] The unique ID of the query execution to stop.}
}
\description{
Stops a query execution. Requires you to have access to the workgroup in
which the query ran.
}
\details{
For code samples using the AWS SDK for Java, see \href{https://docs.aws.amazon.com/athena/latest/ug/code-samples.html}{Examples and Code Samples}
in the \emph{Amazon Athena User Guide}.
}
\section{Request syntax}{
\preformatted{svc$stop_query_execution(
  QueryExecutionId = "string"
)
}
}

\keyword{internal}
