% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_batch_associate_assessment_report_evidence}
\alias{auditmanager_batch_associate_assessment_report_evidence}
\title{Associates a list of evidence to an assessment report in an AWS Audit
Manager assessment}
\usage{
auditmanager_batch_associate_assessment_report_evidence(assessmentId,
  evidenceFolderId, evidenceIds)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the specified assessment.}

\item{evidenceFolderId}{[required] The identifier for the folder in which the evidence is stored.}

\item{evidenceIds}{[required] The list of evidence identifiers.}
}
\description{
Associates a list of evidence to an assessment report in an AWS Audit
Manager assessment.
}
\section{Request syntax}{
\preformatted{svc$batch_associate_assessment_report_evidence(
  assessmentId = "string",
  evidenceFolderId = "string",
  evidenceIds = list(
    "string"
  )
)
}
}

\keyword{internal}
