% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_batch_import_evidence_to_assessment_control}
\alias{auditmanager_batch_import_evidence_to_assessment_control}
\title{Uploads one or more pieces of evidence to the specified control in the
assessment in AWS Audit Manager}
\usage{
auditmanager_batch_import_evidence_to_assessment_control(assessmentId,
  controlSetId, controlId, manualEvidence)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{controlSetId}{[required] The identifier for the specified control set.}

\item{controlId}{[required] The identifier for the specified control.}

\item{manualEvidence}{[required] The list of manual evidence objects.}
}
\description{
Uploads one or more pieces of evidence to the specified control in the
assessment in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$batch_import_evidence_to_assessment_control(
  assessmentId = "string",
  controlSetId = "string",
  controlId = "string",
  manualEvidence = list(
    list(
      s3ResourcePath = "string"
    )
  )
)
}
}

\keyword{internal}
