% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_create_assessment}
\alias{auditmanager_create_assessment}
\title{Creates an assessment in AWS Audit Manager}
\usage{
auditmanager_create_assessment(name, description,
  assessmentReportsDestination, scope, roles, frameworkId, tags)
}
\arguments{
\item{name}{[required] The name of the assessment to be created.}

\item{description}{The optional description of the assessment to be created.}

\item{assessmentReportsDestination}{[required] The assessment report storage destination for the specified assessment
that is being created.}

\item{scope}{[required]}

\item{roles}{[required] The list of roles for the specified assessment.}

\item{frameworkId}{[required] The identifier for the specified framework.}

\item{tags}{The tags associated with the assessment.}
}
\description{
Creates an assessment in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$create_assessment(
  name = "string",
  description = "string",
  assessmentReportsDestination = list(
    destinationType = "S3",
    destination = "string"
  ),
  scope = list(
    awsAccounts = list(
      list(
        id = "string",
        emailAddress = "string",
        name = "string"
      )
    ),
    awsServices = list(
      list(
        serviceName = "string"
      )
    )
  ),
  roles = list(
    list(
      roleType = "PROCESS_OWNER"|"RESOURCE_OWNER",
      roleArn = "string"
    )
  ),
  frameworkId = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
