% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_create_assessment_framework}
\alias{auditmanager_create_assessment_framework}
\title{Creates a custom framework in AWS Audit Manager}
\usage{
auditmanager_create_assessment_framework(name, description,
  complianceType, controlSets)
}
\arguments{
\item{name}{[required] The name of the new custom framework.}

\item{description}{An optional description for the new custom framework.}

\item{complianceType}{The compliance type that the new custom framework supports, such as CIS
or HIPAA.}

\item{controlSets}{[required] The control sets to be associated with the framework.}
}
\description{
Creates a custom framework in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$create_assessment_framework(
  name = "string",
  description = "string",
  complianceType = "string",
  controlSets = list(
    list(
      name = "string",
      controls = list(
        list(
          id = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
