% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_create_assessment_report}
\alias{auditmanager_create_assessment_report}
\title{Creates an assessment report for the specified assessment}
\usage{
auditmanager_create_assessment_report(name, description, assessmentId)
}
\arguments{
\item{name}{[required] The name of the new assessment report.}

\item{description}{The description of the assessment report.}

\item{assessmentId}{[required] The identifier for the specified assessment.}
}
\description{
Creates an assessment report for the specified assessment.
}
\section{Request syntax}{
\preformatted{svc$create_assessment_report(
  name = "string",
  description = "string",
  assessmentId = "string"
)
}
}

\keyword{internal}
