% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_delete_assessment}
\alias{auditmanager_delete_assessment}
\title{Deletes an assessment in AWS Audit Manager}
\usage{
auditmanager_delete_assessment(assessmentId)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}
}
\description{
Deletes an assessment in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$delete_assessment(
  assessmentId = "string"
)
}
}

\keyword{internal}
