% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_delete_assessment_framework}
\alias{auditmanager_delete_assessment_framework}
\title{Deletes a custom framework in AWS Audit Manager}
\usage{
auditmanager_delete_assessment_framework(frameworkId)
}
\arguments{
\item{frameworkId}{[required] The identifier for the specified framework.}
}
\description{
Deletes a custom framework in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$delete_assessment_framework(
  frameworkId = "string"
)
}
}

\keyword{internal}
