% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_delete_assessment_report}
\alias{auditmanager_delete_assessment_report}
\title{Deletes an assessment report from an assessment in AWS Audit Manager}
\usage{
auditmanager_delete_assessment_report(assessmentId, assessmentReportId)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{assessmentReportId}{[required] The unique identifier for the assessment report.}
}
\description{
Deletes an assessment report from an assessment in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$delete_assessment_report(
  assessmentId = "string",
  assessmentReportId = "string"
)
}
}

\keyword{internal}
