% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_deregister_account}
\alias{auditmanager_deregister_account}
\title{Deregisters an account in AWS Audit Manager}
\usage{
auditmanager_deregister_account()
}
\description{
Deregisters an account in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$deregister_account()
}
}

\keyword{internal}
