% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_deregister_organization_admin_account}
\alias{auditmanager_deregister_organization_admin_account}
\title{Deregisters the delegated AWS administrator account from the AWS
organization}
\usage{
auditmanager_deregister_organization_admin_account(adminAccountId)
}
\arguments{
\item{adminAccountId}{The identifier for the specified administrator account.}
}
\description{
Deregisters the delegated AWS administrator account from the AWS
organization.
}
\section{Request syntax}{
\preformatted{svc$deregister_organization_admin_account(
  adminAccountId = "string"
)
}
}

\keyword{internal}
