% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_account_status}
\alias{auditmanager_get_account_status}
\title{Returns the registration status of an account in AWS Audit Manager}
\usage{
auditmanager_get_account_status()
}
\description{
Returns the registration status of an account in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$get_account_status()
}
}

\keyword{internal}
