% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_assessment_framework}
\alias{auditmanager_get_assessment_framework}
\title{Returns a framework from AWS Audit Manager}
\usage{
auditmanager_get_assessment_framework(frameworkId)
}
\arguments{
\item{frameworkId}{[required] The identifier for the specified framework.}
}
\description{
Returns a framework from AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$get_assessment_framework(
  frameworkId = "string"
)
}
}

\keyword{internal}
