% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_control}
\alias{auditmanager_get_control}
\title{Returns a control from AWS Audit Manager}
\usage{
auditmanager_get_control(controlId)
}
\arguments{
\item{controlId}{[required] The identifier for the specified control.}
}
\description{
Returns a control from AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$get_control(
  controlId = "string"
)
}
}

\keyword{internal}
