% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_delegations}
\alias{auditmanager_get_delegations}
\title{Returns a list of delegations from an audit owner to a delegate}
\usage{
auditmanager_get_delegations(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns a list of delegations from an audit owner to a delegate.
}
\section{Request syntax}{
\preformatted{svc$get_delegations(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
