% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_evidence}
\alias{auditmanager_get_evidence}
\title{Returns evidence from AWS Audit Manager}
\usage{
auditmanager_get_evidence(assessmentId, controlSetId, evidenceFolderId,
  evidenceId)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{controlSetId}{[required] The identifier for the specified control set.}

\item{evidenceFolderId}{[required] The identifier for the folder in which the evidence is stored.}

\item{evidenceId}{[required] The identifier for the evidence.}
}
\description{
Returns evidence from AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$get_evidence(
  assessmentId = "string",
  controlSetId = "string",
  evidenceFolderId = "string",
  evidenceId = "string"
)
}
}

\keyword{internal}
