% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_evidence_folders_by_assessment}
\alias{auditmanager_get_evidence_folders_by_assessment}
\title{Returns the evidence folders from a specified assessment in AWS Audit
Manager}
\usage{
auditmanager_get_evidence_folders_by_assessment(assessmentId, nextToken,
  maxResults)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns the evidence folders from a specified assessment in AWS Audit
Manager.
}
\section{Request syntax}{
\preformatted{svc$get_evidence_folders_by_assessment(
  assessmentId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
