% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_services_in_scope}
\alias{auditmanager_get_services_in_scope}
\title{Returns a list of the in-scope AWS services for the specified assessment}
\usage{
auditmanager_get_services_in_scope()
}
\description{
Returns a list of the in-scope AWS services for the specified
assessment.
}
\section{Request syntax}{
\preformatted{svc$get_services_in_scope()
}
}

\keyword{internal}
