% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_settings}
\alias{auditmanager_get_settings}
\title{Returns the settings for the specified AWS account}
\usage{
auditmanager_get_settings(attribute)
}
\arguments{
\item{attribute}{[required] The list of \code{SettingAttribute} enum values.}
}
\description{
Returns the settings for the specified AWS account.
}
\section{Request syntax}{
\preformatted{svc$get_settings(
  attribute = "ALL"|"IS_AWS_ORG_ENABLED"|"SNS_TOPIC"|"DEFAULT_ASSESSMENT_REPORTS_DESTINATION"|"DEFAULT_PROCESS_OWNERS"
)
}
}

\keyword{internal}
