% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_assessment_reports}
\alias{auditmanager_list_assessment_reports}
\title{Returns a list of assessment reports created in AWS Audit Manager}
\usage{
auditmanager_list_assessment_reports(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns a list of assessment reports created in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$list_assessment_reports(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
