% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_assessments}
\alias{auditmanager_list_assessments}
\title{Returns a list of current and past assessments from AWS Audit Manager}
\usage{
auditmanager_list_assessments(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns a list of current and past assessments from AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$list_assessments(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
