% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_controls}
\alias{auditmanager_list_controls}
\title{Returns a list of controls from AWS Audit Manager}
\usage{
auditmanager_list_controls(controlType, nextToken, maxResults)
}
\arguments{
\item{controlType}{[required] The type of control, such as standard or custom.}

\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns a list of controls from AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$list_controls(
  controlType = "Standard"|"Custom",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
