% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_keywords_for_data_source}
\alias{auditmanager_list_keywords_for_data_source}
\title{Returns a list of keywords that pre-mapped to the specified control data
source}
\usage{
auditmanager_list_keywords_for_data_source(source, nextToken,
  maxResults)
}
\arguments{
\item{source}{[required] The control mapping data source to which the keywords apply.}

\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns a list of keywords that pre-mapped to the specified control data
source.
}
\section{Request syntax}{
\preformatted{svc$list_keywords_for_data_source(
  source = "AWS_Cloudtrail"|"AWS_Config"|"AWS_Security_Hub"|"AWS_API_Call"|"MANUAL",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
