% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_notifications}
\alias{auditmanager_list_notifications}
\title{Returns a list of all AWS Audit Manager notifications}
\usage{
auditmanager_list_notifications(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns a list of all AWS Audit Manager notifications.
}
\section{Request syntax}{
\preformatted{svc$list_notifications(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
