% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_register_account}
\alias{auditmanager_register_account}
\title{Enables AWS Audit Manager for the specified AWS account}
\usage{
auditmanager_register_account(kmsKey, delegatedAdminAccount)
}
\arguments{
\item{kmsKey}{The AWS KMS key details.}

\item{delegatedAdminAccount}{The delegated administrator account for AWS Audit Manager.}
}
\description{
Enables AWS Audit Manager for the specified AWS account.
}
\section{Request syntax}{
\preformatted{svc$register_account(
  kmsKey = "string",
  delegatedAdminAccount = "string"
)
}
}

\keyword{internal}
