% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_register_organization_admin_account}
\alias{auditmanager_register_organization_admin_account}
\title{Enables an AWS account within the organization as the delegated
administrator for AWS Audit Manager}
\usage{
auditmanager_register_organization_admin_account(adminAccountId)
}
\arguments{
\item{adminAccountId}{[required] The identifier for the specified delegated administrator account.}
}
\description{
Enables an AWS account within the organization as the delegated
administrator for AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$register_organization_admin_account(
  adminAccountId = "string"
)
}
}

\keyword{internal}
