% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_untag_resource}
\alias{auditmanager_untag_resource}
\title{Removes a tag from a resource in AWS Audit Manager}
\usage{
auditmanager_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the specified resource.}

\item{tagKeys}{[required] The name or key of the tag.}
}
\description{
Removes a tag from a resource in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
