% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_assessment_control_set_status}
\alias{auditmanager_update_assessment_control_set_status}
\title{Updates the status of a control set in an AWS Audit Manager assessment}
\usage{
auditmanager_update_assessment_control_set_status(assessmentId,
  controlSetId, status, comment)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{controlSetId}{[required] The identifier for the specified control set.}

\item{status}{[required] The status of the control set that is being updated.}

\item{comment}{[required] The comment related to the status update.}
}
\description{
Updates the status of a control set in an AWS Audit Manager assessment.
}
\section{Request syntax}{
\preformatted{svc$update_assessment_control_set_status(
  assessmentId = "string",
  controlSetId = "string",
  status = "ACTIVE"|"UNDER_REVIEW"|"REVIEWED",
  comment = "string"
)
}
}

\keyword{internal}
