% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_assessment_framework}
\alias{auditmanager_update_assessment_framework}
\title{Updates a custom framework in AWS Audit Manager}
\usage{
auditmanager_update_assessment_framework(frameworkId, name, description,
  complianceType, controlSets)
}
\arguments{
\item{frameworkId}{[required] The identifier for the specified framework.}

\item{name}{[required] The name of the framework to be updated.}

\item{description}{The description of the framework that is to be updated.}

\item{complianceType}{The compliance type that the new custom framework supports, such as CIS
or HIPAA.}

\item{controlSets}{[required] The control sets associated with the framework.}
}
\description{
Updates a custom framework in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$update_assessment_framework(
  frameworkId = "string",
  name = "string",
  description = "string",
  complianceType = "string",
  controlSets = list(
    list(
      id = "string",
      name = "string",
      controls = list(
        list(
          id = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
