% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_control}
\alias{auditmanager_update_control}
\title{Updates a custom control in AWS Audit Manager}
\usage{
auditmanager_update_control(controlId, name, description,
  testingInformation, actionPlanTitle, actionPlanInstructions,
  controlMappingSources)
}
\arguments{
\item{controlId}{[required] The identifier for the specified control.}

\item{name}{[required] The name of the control to be updated.}

\item{description}{The optional description of the control.}

\item{testingInformation}{The steps that to follow to determine if the control has been satisfied.}

\item{actionPlanTitle}{The title of the action plan for remediating the control.}

\item{actionPlanInstructions}{The recommended actions to carry out if the control is not fulfilled.}

\item{controlMappingSources}{[required] The data source that determines from where AWS Audit Manager collects
evidence for the control.}
}
\description{
Updates a custom control in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$update_control(
  controlId = "string",
  name = "string",
  description = "string",
  testingInformation = "string",
  actionPlanTitle = "string",
  actionPlanInstructions = "string",
  controlMappingSources = list(
    list(
      sourceId = "string",
      sourceName = "string",
      sourceDescription = "string",
      sourceSetUpOption = "System_Controls_Mapping"|"Procedural_Controls_Mapping",
      sourceType = "AWS_Cloudtrail"|"AWS_Config"|"AWS_Security_Hub"|"AWS_API_Call"|"MANUAL",
      sourceKeyword = list(
        keywordInputType = "SELECT_FROM_LIST",
        keywordValue = "string"
      ),
      sourceFrequency = "DAILY"|"WEEKLY"|"MONTHLY",
      troubleshootingText = "string"
    )
  )
)
}
}

\keyword{internal}
