% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedairuntime_operations.R
\name{augmentedairuntime_describe_human_loop}
\alias{augmentedairuntime_describe_human_loop}
\title{Returns information about the specified human loop}
\usage{
augmentedairuntime_describe_human_loop(HumanLoopName)
}
\arguments{
\item{HumanLoopName}{[required] The name of the human loop.}
}
\description{
Returns information about the specified human loop.
}
\section{Request syntax}{
\preformatted{svc$describe_human_loop(
  HumanLoopName = "string"
)
}
}

\keyword{internal}
