% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedairuntime_operations.R
\name{augmentedairuntime_list_human_loops}
\alias{augmentedairuntime_list_human_loops}
\title{Returns information about human loops, given the specified parameters}
\usage{
augmentedairuntime_list_human_loops(CreationTimeAfter,
  CreationTimeBefore, SortOrder, NextToken, MaxResults)
}
\arguments{
\item{CreationTimeAfter}{(Optional) The timestamp of the date when you want the human loops to
begin. For example, \code{1551000000}.}

\item{CreationTimeBefore}{(Optional) The timestamp of the date before which you want the human
loops to begin. For example, \code{1550000000}.}

\item{SortOrder}{An optional value that specifies whether you want the results sorted in
\code{Ascending} or \code{Descending} order.}

\item{NextToken}{A token to resume pagination.}

\item{MaxResults}{The total number of items to return. If the total number of available
items is more than the value specified in \code{MaxResults}, then a
\code{NextToken} will be provided in the output that you can use to resume
pagination.}
}
\description{
Returns information about human loops, given the specified parameters.
}
\section{Request syntax}{
\preformatted{svc$list_human_loops(
  CreationTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  CreationTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  SortOrder = "Ascending"|"Descending",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
