% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedairuntime_operations.R
\name{augmentedairuntime_start_human_loop}
\alias{augmentedairuntime_start_human_loop}
\title{Starts a human loop, provided that at least one activation condition is
met}
\usage{
augmentedairuntime_start_human_loop(HumanLoopName, FlowDefinitionArn,
  HumanLoopInput, DataAttributes)
}
\arguments{
\item{HumanLoopName}{[required] The name of the human loop.}

\item{FlowDefinitionArn}{[required] The Amazon Resource Name (ARN) of the flow definition.}

\item{HumanLoopInput}{[required] An object containing information about the human loop.}

\item{DataAttributes}{Attributes of the data specified by the customer.}
}
\description{
Starts a human loop, provided that at least one activation condition is
met.
}
\section{Request syntax}{
\preformatted{svc$start_human_loop(
  HumanLoopName = "string",
  FlowDefinitionArn = "string",
  HumanLoopInput = list(
    InputContent = "string"
  ),
  DataAttributes = list(
    ContentClassifiers = list(
      "FreeOfPersonallyIdentifiableInformation"|"FreeOfAdultContent"
    )
  )
)
}
}

\keyword{internal}
