% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedairuntime_operations.R
\name{augmentedairuntime_stop_human_loop}
\alias{augmentedairuntime_stop_human_loop}
\title{Stops the specified human loop}
\usage{
augmentedairuntime_stop_human_loop(HumanLoopName)
}
\arguments{
\item{HumanLoopName}{[required] The name of the human loop that you want to stop.}
}
\description{
Stops the specified human loop.
}
\section{Request syntax}{
\preformatted{svc$stop_human_loop(
  HumanLoopName = "string"
)
}
}

\keyword{internal}
