% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_delete_launch_configuration}
\alias{autoscaling_delete_launch_configuration}
\title{Deletes the specified launch configuration}
\usage{
autoscaling_delete_launch_configuration(LaunchConfigurationName)
}
\arguments{
\item{LaunchConfigurationName}{[required] The name of the launch configuration.}
}
\description{
Deletes the specified launch configuration.
}
\details{
The launch configuration must not be attached to an Auto Scaling group.
When this call completes, the launch configuration is no longer
available for use.
}
\section{Request syntax}{
\preformatted{svc$delete_launch_configuration(
  LaunchConfigurationName = "string"
)
}
}

\section{Examples}{
\preformatted{# This example deletes the specified launch configuration.
svc$delete_launch_configuration(
  LaunchConfigurationName = "my-launch-config"
)
}
}

\keyword{internal}
