% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_adjustment_types}
\alias{autoscaling_describe_adjustment_types}
\title{Describes the policy adjustment types for use with PutScalingPolicy}
\usage{
autoscaling_describe_adjustment_types()
}
\description{
Describes the policy adjustment types for use with PutScalingPolicy.
}
\examples{
\dontrun{
# This example describes the available adjustment types.
svc$describe_adjustment_types()
}

}
\keyword{internal}
