% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_launch_configurations}
\alias{autoscaling_describe_launch_configurations}
\title{Describes one or more launch configurations}
\usage{
autoscaling_describe_launch_configurations(LaunchConfigurationNames,
  NextToken, MaxRecords)
}
\arguments{
\item{LaunchConfigurationNames}{The launch configuration names. If you omit this parameter, all launch
configurations are described.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxRecords}{The maximum number of items to return with this call. The default value
is \code{50} and the maximum value is \code{100}.}
}
\description{
Describes one or more launch configurations.
}
\section{Request syntax}{
\preformatted{svc$describe_launch_configurations(
  LaunchConfigurationNames = list(
    "string"
  ),
  NextToken = "string",
  MaxRecords = 123
)
}
}

\section{Examples}{
\preformatted{# This example describes the specified launch configuration.
svc$describe_launch_configurations(
  LaunchConfigurationNames = list(
    "my-launch-config"
  )
)
}
}

\keyword{internal}
