% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_lifecycle_hook_types}
\alias{autoscaling_describe_lifecycle_hook_types}
\title{Describes the available types of lifecycle hooks}
\usage{
autoscaling_describe_lifecycle_hook_types()
}
\description{
Describes the available types of lifecycle hooks.
}
\details{
The following hook types are supported:
\itemize{
\item autoscaling:EC2\\_INSTANCE\\_LAUNCHING
\item autoscaling:EC2\\_INSTANCE\\_TERMINATING
}
}
\section{Examples}{
\preformatted{# This example describes the available lifecycle hook types.
svc$describe_lifecycle_hook_types()
}
}

\keyword{internal}
