% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_scaling_process_types}
\alias{autoscaling_describe_scaling_process_types}
\title{Describes the scaling process types for use with ResumeProcesses and
SuspendProcesses}
\usage{
autoscaling_describe_scaling_process_types()
}
\description{
Describes the scaling process types for use with ResumeProcesses and
SuspendProcesses.
}
\section{Examples}{
\preformatted{# This example describes the Auto Scaling process types.
svc$describe_scaling_process_types()
}
}

\keyword{internal}
