% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_global_settings}
\alias{backup_describe_global_settings}
\title{The current feature settings for the AWS Account}
\usage{
backup_describe_global_settings()
}
\description{
The current feature settings for the AWS Account.
}
\section{Request syntax}{
\preformatted{svc$describe_global_settings()
}
}

\keyword{internal}
