% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_update_global_settings}
\alias{backup_update_global_settings}
\title{Updates the current global settings for the AWS Account}
\usage{
backup_update_global_settings(GlobalSettings)
}
\arguments{
\item{GlobalSettings}{A list of resources along with the opt-in preferences for the account.}
}
\description{
Updates the current global settings for the AWS Account. Use the
\code{DescribeGlobalSettings} API to determine the current settings.
}
\section{Request syntax}{
\preformatted{svc$update_global_settings(
  GlobalSettings = list(
    "string"
  )
)
}
}

\keyword{internal}
