% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_update_region_settings}
\alias{backup_update_region_settings}
\title{Updates the current service opt-in settings for the Region}
\usage{
backup_update_region_settings(ResourceTypeOptInPreference)
}
\arguments{
\item{ResourceTypeOptInPreference}{Updates the list of services along with the opt-in preferences for the
region.}
}
\description{
Updates the current service opt-in settings for the Region. If the
service has a value set to \code{true}, AWS Backup attempts to protect that
service's resources in this Region, when included in an on-demand backup
or scheduled backup plan. If the value is set to \code{false} for a service,
AWS Backup does not attempt to protect that service's resources in this
Region.
}
\section{Request syntax}{
\preformatted{svc$update_region_settings(
  ResourceTypeOptInPreference = list(
    TRUE|FALSE
  )
)
}
}

\keyword{internal}
