% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{bedrockdataautomation}
\alias{bedrockdataautomation}
\title{Data Automation for Amazon Bedrock}
\usage{
bedrockdataautomation(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
\item{\strong{use_dual_stack}: Set this to \code{true} to use the dualstack (IPv4 and IPv6) endpoint for a service, where available, falling back to the regular endpoint when it isn't. Defaults to the \code{AWS_USE_DUALSTACK_ENDPOINT} environment variable when unset.}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Amazon Bedrock Data Automation BuildTime
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- bedrockdataautomation(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string",
    use_dual_stack = "logical"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.machine.learning:bedrockdataautomation_copy_blueprint_stage]{copy_blueprint_stage} \tab Copies a Blueprint from one stage to another\cr
\link[paws.machine.learning:bedrockdataautomation_create_blueprint]{create_blueprint} \tab Creates an Amazon Bedrock Data Automation Blueprint\cr
\link[paws.machine.learning:bedrockdataautomation_create_blueprint_version]{create_blueprint_version} \tab Creates a new version of an existing Amazon Bedrock Data Automation Blueprint\cr
\link[paws.machine.learning:bedrockdataautomation_create_data_automation_library]{create_data_automation_library} \tab Creates an Amazon Bedrock Data Automation Library\cr
\link[paws.machine.learning:bedrockdataautomation_create_data_automation_project]{create_data_automation_project} \tab Creates an Amazon Bedrock Data Automation Project\cr
\link[paws.machine.learning:bedrockdataautomation_delete_blueprint]{delete_blueprint} \tab Deletes an existing Amazon Bedrock Data Automation Blueprint\cr
\link[paws.machine.learning:bedrockdataautomation_delete_data_automation_library]{delete_data_automation_library} \tab Deletes an existing Amazon Bedrock Data Automation Library\cr
\link[paws.machine.learning:bedrockdataautomation_delete_data_automation_project]{delete_data_automation_project} \tab Deletes an existing Amazon Bedrock Data Automation Project\cr
\link[paws.machine.learning:bedrockdataautomation_get_blueprint]{get_blueprint} \tab Gets an existing Amazon Bedrock Data Automation Blueprint\cr
\link[paws.machine.learning:bedrockdataautomation_get_blueprint_optimization_status]{get_blueprint_optimization_status} \tab API used to get blueprint optimization status\cr
\link[paws.machine.learning:bedrockdataautomation_get_data_automation_library]{get_data_automation_library} \tab Gets an existing Amazon Bedrock Data Automation Library\cr
\link[paws.machine.learning:bedrockdataautomation_get_data_automation_library_entity]{get_data_automation_library_entity} \tab Gets an existing entity based on entity type from the library\cr
\link[paws.machine.learning:bedrockdataautomation_get_data_automation_library_ingestion_job]{get_data_automation_library_ingestion_job} \tab API used to get status of data automation library ingestion job\cr
\link[paws.machine.learning:bedrockdataautomation_get_data_automation_project]{get_data_automation_project} \tab Gets an existing Amazon Bedrock Data Automation Project\cr
\link[paws.machine.learning:bedrockdataautomation_invoke_blueprint_optimization_async]{invoke_blueprint_optimization_async} \tab Invoke an async job to perform Blueprint Optimization\cr
\link[paws.machine.learning:bedrockdataautomation_invoke_data_automation_library_ingestion_job]{invoke_data_automation_library_ingestion_job} \tab Async API: Invoke data automation library ingestion job\cr
\link[paws.machine.learning:bedrockdataautomation_list_blueprints]{list_blueprints} \tab Lists all existing Amazon Bedrock Data Automation Blueprints\cr
\link[paws.machine.learning:bedrockdataautomation_list_data_automation_libraries]{list_data_automation_libraries} \tab Lists all existing Amazon Bedrock Data Automation Libraries\cr
\link[paws.machine.learning:bedrockdataautomation_list_data_automation_library_entities]{list_data_automation_library_entities} \tab Lists all stored entities in the library\cr
\link[paws.machine.learning:bedrockdataautomation_list_data_automation_library_ingestion_jobs]{list_data_automation_library_ingestion_jobs} \tab Lists all data automation library ingestion jobs\cr
\link[paws.machine.learning:bedrockdataautomation_list_data_automation_projects]{list_data_automation_projects} \tab Lists all existing Amazon Bedrock Data Automation Projects\cr
\link[paws.machine.learning:bedrockdataautomation_list_tags_for_resource]{list_tags_for_resource} \tab List tags for an Amazon Bedrock Data Automation resource\cr
\link[paws.machine.learning:bedrockdataautomation_tag_resource]{tag_resource} \tab Tag an Amazon Bedrock Data Automation resource\cr
\link[paws.machine.learning:bedrockdataautomation_untag_resource]{untag_resource} \tab Untag an Amazon Bedrock Data Automation resource\cr
\link[paws.machine.learning:bedrockdataautomation_update_blueprint]{update_blueprint} \tab Updates an existing Amazon Bedrock Data Automation Blueprint\cr
\link[paws.machine.learning:bedrockdataautomation_update_data_automation_library]{update_data_automation_library} \tab Updates an existing Amazon Bedrock Data Automation Library\cr
\link[paws.machine.learning:bedrockdataautomation_update_data_automation_project]{update_data_automation_project} \tab Updates an existing Amazon Bedrock Data Automation Project
}
}

\examples{
\dontrun{
svc <- bedrockdataautomation()
svc$copy_blueprint_stage(
  Foo = 123
)
}

}
