% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_get_device}
\alias{braket_get_device}
\title{Retrieves the devices available in Amazon Braket}
\usage{
braket_get_device(deviceArn)
}
\arguments{
\item{deviceArn}{[required] The ARN of the device to retrieve.}
}
\description{
Retrieves the devices available in Amazon Braket.
}
\section{Request syntax}{
\preformatted{svc$get_device(
  deviceArn = "string"
)
}
}

\keyword{internal}
