% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_search_quantum_tasks}
\alias{braket_search_quantum_tasks}
\title{Searches for tasks that match the specified filter values}
\usage{
braket_search_quantum_tasks(filters, maxResults, nextToken)
}
\arguments{
\item{filters}{[required] Array of \code{SearchQuantumTasksFilter} objects.}

\item{maxResults}{Maximum number of results to return in the response.}

\item{nextToken}{A token used for pagination of results returned in the response. Use the
token returned from the previous request continue results where the
previous request ended.}
}
\description{
Searches for tasks that match the specified filter values.
}
\section{Request syntax}{
\preformatted{svc$search_quantum_tasks(
  filters = list(
    list(
      name = "string",
      operator = "LT"|"LTE"|"EQUAL"|"GT"|"GTE"|"BETWEEN",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
