% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_service.R
\name{chime}
\alias{chime}
\title{Amazon Chime}
\usage{
chime()
}
\description{
The Amazon Chime API (application programming interface) is designed for
administrators to use to perform key tasks, such as creating and
managing Amazon Chime accounts and users. This guide provides detailed
information about the Amazon Chime API, including operations, types,
inputs and outputs, and error codes.

You can use an AWS SDK, the AWS Command Line Interface (AWS CLI), or the
REST API to make API calls. We recommend using an AWS SDK or the AWS
CLI. Each API operation includes links to information about using it
with a language-specific AWS SDK or the AWS CLI.

Using an AWS SDK

:   You don't need to write code to calculate a signature for request
authentication. The SDK clients authenticate your requests by using
access keys that you provide. For more information about AWS SDKs,
see the \href{http://aws.amazon.com/developer/}{AWS Developer Center}.

Using the AWS CLI

:   Use your access keys with the AWS CLI to make API calls. For
information about setting up the AWS CLI, see \href{https://docs.aws.amazon.com/cli/latest/userguide/installing.html}{Installing the AWS Command Line Interface}
in the \emph{AWS Command Line Interface User Guide}. For a list of
available Amazon Chime commands, see the \href{https://docs.aws.amazon.com/cli/latest/reference/chime/index.html}{Amazon Chime commands}
in the \emph{AWS CLI Command Reference}.

Using REST API

:   If you use REST to make API calls, you must authenticate your
request by providing a signature. Amazon Chime supports signature
version 4. For more information, see \href{https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html}{Signature Version 4 Signing Process}
in the \emph{Amazon Web Services General Reference}.\preformatted{When making REST API calls, use the service name `chime` and REST
endpoint `https://service.chime.aws.amazon.com`.
}

Administrative permissions are controlled using AWS Identity and Access
Management (IAM). For more information, see \href{https://docs.aws.amazon.com/chime/latest/ag/control-access.html}{Control Access to the Amazon Chime Console}
in the \emph{Amazon Chime Administration Guide}.
}
\section{Example}{
\preformatted{chime <- paws::chime()
chime$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=chime_associate_phone_number_with_user]{associate_phone_number_with_user} \tab Associates a phone number with the specified Amazon Chime user \cr
\link[=chime_associate_phone_numbers_with_voice_connector]{associate_phone_numbers_with_voice_connector} \tab Associates a phone number with the specified Amazon Chime Voice Connector \cr
\link[=chime_batch_delete_phone_number]{batch_delete_phone_number} \tab Moves phone numbers into the DELETION QUEUE \cr
\link[=chime_batch_suspend_user]{batch_suspend_user} \tab Suspends up to 50 users from a Team or EnterpriseLWA Amazon Chime account \cr
\link[=chime_batch_unsuspend_user]{batch_unsuspend_user} \tab Removes the suspension from up to 50 previously suspended users for the specified Amazon Chime EnterpriseLWA account \cr
\link[=chime_batch_update_phone_number]{batch_update_phone_number} \tab Updates phone number product types \cr
\link[=chime_batch_update_user]{batch_update_user} \tab Updates user details within the UpdateUserRequestItem object for up to 20 users for the specified Amazon Chime account \cr
\link[=chime_create_account]{create_account} \tab Creates an Amazon Chime account under the administrator's AWS account \cr
\link[=chime_create_phone_number_order]{create_phone_number_order} \tab Creates an order for phone numbers to be provisioned \cr
\link[=chime_create_voice_connector]{create_voice_connector} \tab Creates an Amazon Chime Voice Connector under the administrator's AWS account \cr
\link[=chime_delete_account]{delete_account} \tab Deletes the specified Amazon Chime account \cr
\link[=chime_delete_phone_number]{delete_phone_number} \tab Moves the specified phone number into the DELETION QUEUE \cr
\link[=chime_delete_voice_connector]{delete_voice_connector} \tab Deletes the specified Amazon Chime Voice Connector \cr
\link[=chime_delete_voice_connector_origination]{delete_voice_connector_origination} \tab Deletes the origination settings for the specified Amazon Chime Voice Connector \cr
\link[=chime_delete_voice_connector_termination]{delete_voice_connector_termination} \tab Deletes the termination settings for the specified Amazon Chime Voice Connector \cr
\link[=chime_delete_voice_connector_termination_credentials]{delete_voice_connector_termination_credentials} \tab Deletes the specified SIP credentials used by your equipment to authenticate during call termination \cr
\link[=chime_disassociate_phone_number_from_user]{disassociate_phone_number_from_user} \tab Disassociates the primary provisioned phone number from the specified Amazon Chime user \cr
\link[=chime_disassociate_phone_numbers_from_voice_connector]{disassociate_phone_numbers_from_voice_connector} \tab Disassociates the specified phone number from the specified Amazon Chime Voice Connector \cr
\link[=chime_get_account]{get_account} \tab Retrieves details for the specified Amazon Chime account, such as account type and supported licenses \cr
\link[=chime_get_account_settings]{get_account_settings} \tab Retrieves account settings for the specified Amazon Chime account ID, such as remote control and dial out settings \cr
\link[=chime_get_global_settings]{get_global_settings} \tab Retrieves global settings for the administrator's AWS account, such as Amazon Chime Business Calling and Amazon Chime Voice Connector settings \cr
\link[=chime_get_phone_number]{get_phone_number} \tab Retrieves details for the specified phone number ID, such as associations, capabilities, and product type \cr
\link[=chime_get_phone_number_order]{get_phone_number_order} \tab Retrieves details for the specified phone number order, such as order creation timestamp, phone numbers in E \cr
\link[=chime_get_user]{get_user} \tab Retrieves details for the specified user ID, such as primary email address, license type, and personal meeting PIN \cr
\link[=chime_get_user_settings]{get_user_settings} \tab Retrieves settings for the specified user ID, such as any associated phone number settings \cr
\link[=chime_get_voice_connector]{get_voice_connector} \tab Retrieves details for the specified Amazon Chime Voice Connector, such as timestamps, name, outbound host, and encryption requirements \cr
\link[=chime_get_voice_connector_origination]{get_voice_connector_origination} \tab Retrieves origination setting details for the specified Amazon Chime Voice Connector \cr
\link[=chime_get_voice_connector_termination]{get_voice_connector_termination} \tab Retrieves termination setting details for the specified Amazon Chime Voice Connector \cr
\link[=chime_get_voice_connector_termination_health]{get_voice_connector_termination_health} \tab Retrieves information about the last time a SIP OPTIONS ping was received from your SIP infrastructure for the specified Amazon Chime Voice Connector\cr
\link[=chime_invite_users]{invite_users} \tab Sends email invites to as many as 50 users, inviting them to the specified Amazon Chime Team account \cr
\link[=chime_list_accounts]{list_accounts} \tab Lists the Amazon Chime accounts under the administrator's AWS account \cr
\link[=chime_list_phone_number_orders]{list_phone_number_orders} \tab Lists the phone number orders for the administrator's Amazon Chime account \cr
\link[=chime_list_phone_numbers]{list_phone_numbers} \tab Lists the phone numbers for the specified Amazon Chime account, Amazon Chime user, or Amazon Chime Voice Connector \cr
\link[=chime_list_users]{list_users} \tab Lists the users that belong to the specified Amazon Chime account \cr
\link[=chime_list_voice_connector_termination_credentials]{list_voice_connector_termination_credentials} \tab Lists the SIP credentials for the specified Amazon Chime Voice Connector \cr
\link[=chime_list_voice_connectors]{list_voice_connectors} \tab Lists the Amazon Chime Voice Connectors for the administrator's AWS account \cr
\link[=chime_logout_user]{logout_user} \tab Logs out the specified user from all of the devices they are currently logged into \cr
\link[=chime_put_voice_connector_origination]{put_voice_connector_origination} \tab Adds origination settings for the specified Amazon Chime Voice Connector \cr
\link[=chime_put_voice_connector_termination]{put_voice_connector_termination} \tab Adds termination settings for the specified Amazon Chime Voice Connector \cr
\link[=chime_put_voice_connector_termination_credentials]{put_voice_connector_termination_credentials} \tab Adds termination SIP credentials for the specified Amazon Chime Voice Connector \cr
\link[=chime_reset_personal_pin]{reset_personal_pin} \tab Resets the personal meeting PIN for the specified user on an Amazon Chime account \cr
\link[=chime_restore_phone_number]{restore_phone_number} \tab Moves a phone number from the DELETION QUEUE back into the phone number INVENTORY \cr
\link[=chime_search_available_phone_numbers]{search_available_phone_numbers} \tab Searches phone numbers that can be ordered \cr
\link[=chime_update_account]{update_account} \tab Updates account details for the specified Amazon Chime account \cr
\link[=chime_update_account_settings]{update_account_settings} \tab Updates the settings for the specified Amazon Chime account \cr
\link[=chime_update_global_settings]{update_global_settings} \tab Updates global settings for the administrator's AWS account, such as Amazon Chime Business Calling and Amazon Chime Voice Connector settings \cr
\link[=chime_update_phone_number]{update_phone_number} \tab Updates phone number details, such as product type, for the specified phone number ID \cr
\link[=chime_update_user]{update_user} \tab Updates user details for a specified user ID \cr
\link[=chime_update_user_settings]{update_user_settings} \tab Updates the settings for the specified user, such as phone number settings \cr
\link[=chime_update_voice_connector]{update_voice_connector} \tab Updates details for the specified Amazon Chime Voice Connector
}
}

