% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_batch_create_attendee}
\alias{chime_batch_create_attendee}
\title{Creates up to 100 new attendees for an active Amazon Chime SDK meeting}
\usage{
chime_batch_create_attendee(MeetingId, Attendees)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{Attendees}{[required] The request containing the attendees to create.}
}
\description{
Creates up to 100 new attendees for an active Amazon Chime SDK meeting.
For more information about the Amazon Chime SDK, see \href{https://docs.aws.amazon.com/chime/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK} in
the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$batch_create_attendee(
  MeetingId = "string",
  Attendees = list(
    list(
      ExternalUserId = "string",
      Tags = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
