% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_batch_delete_phone_number}
\alias{chime_batch_delete_phone_number}
\title{Moves phone numbers into the Deletion queue}
\usage{
chime_batch_delete_phone_number(PhoneNumberIds)
}
\arguments{
\item{PhoneNumberIds}{[required] List of phone number IDs.}
}
\description{
Moves phone numbers into the \strong{Deletion queue}. Phone numbers must be
disassociated from any users or Amazon Chime Voice Connectors before
they can be deleted.

Phone numbers remain in the \strong{Deletion queue} for 7 days before they
are deleted permanently.
}
\section{Request syntax}{
\preformatted{svc$batch_delete_phone_number(
  PhoneNumberIds = list(
    "string"
  )
)
}
}

\keyword{internal}
