% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_batch_update_user}
\alias{chime_batch_update_user}
\title{Updates user details within the UpdateUserRequestItem object for up to
20 users for the specified Amazon Chime account}
\usage{
chime_batch_update_user(AccountId, UpdateUserRequestItems)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UpdateUserRequestItems}{[required] The request containing the user IDs and details to update.}
}
\description{
Updates user details within the UpdateUserRequestItem object for up to
20 users for the specified Amazon Chime account. Currently, only
\code{LicenseType} updates are supported for this action.
}
\section{Request syntax}{
\preformatted{svc$batch_update_user(
  AccountId = "string",
  UpdateUserRequestItems = list(
    list(
      UserId = "string",
      LicenseType = "Basic"|"Plus"|"Pro"|"ProTrial",
      UserType = "PrivateUser"|"SharedDevice",
      AlexaForBusinessMetadata = list(
        IsAlexaForBusinessEnabled = TRUE|FALSE,
        AlexaForBusinessRoomArn = "string"
      )
    )
  )
)
}
}

\keyword{internal}
