% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_bot}
\alias{chime_create_bot}
\title{Creates a bot for an Amazon Chime Enterprise account}
\usage{
chime_create_bot(AccountId, DisplayName, Domain)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{DisplayName}{[required] The bot display name.}

\item{Domain}{The domain of the Amazon Chime Enterprise account.}
}
\description{
Creates a bot for an Amazon Chime Enterprise account.
}
\section{Request syntax}{
\preformatted{svc$create_bot(
  AccountId = "string",
  DisplayName = "string",
  Domain = "string"
)
}
}

\keyword{internal}
