% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_channel}
\alias{chime_create_channel}
\title{Creates a channel to which you can add users and send messages}
\usage{
chime_create_channel(AppInstanceArn, Name, Mode, Privacy, Metadata,
  ClientRequestToken, Tags)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the channel request.}

\item{Name}{[required] The name of the channel.}

\item{Mode}{The channel mode: \code{UNRESTRICTED} or \code{RESTRICTED}. Administrators,
moderators, and channel members can add themselves and other members to
unrestricted channels. Only administrators and moderators can add
members to restricted channels.}

\item{Privacy}{The channel's privacy level: \code{PUBLIC} or \code{PRIVATE}. Private channels
aren't discoverable by users outside the channel. Public channels are
discoverable by anyone in the app instance.}

\item{Metadata}{The metadata of the creation request. Limited to 1KB and UTF-8.}

\item{ClientRequestToken}{[required] The client token for the request. An \code{Idempotency} token.}

\item{Tags}{}
}
\description{
Creates a channel to which you can add users and send messages.

\strong{Restriction}: You can't change a channel's privacy.
}
\section{Request syntax}{
\preformatted{svc$create_channel(
  AppInstanceArn = "string",
  Name = "string",
  Mode = "UNRESTRICTED"|"RESTRICTED",
  Privacy = "PUBLIC"|"PRIVATE",
  Metadata = "string",
  ClientRequestToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
