% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_channel_moderator}
\alias{chime_create_channel_moderator}
\title{Creates a new ChannelModerator}
\usage{
chime_create_channel_moderator(ChannelArn, ChannelModeratorArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{ChannelModeratorArn}{[required] The ARN of the moderator.}
}
\description{
Creates a new \code{ChannelModerator}. A channel moderator can:
}
\details{
\itemize{
\item Add and remove other members of the channel.
\item Add and remove other moderators of the channel.
\item Add and remove user bans for the channel.
\item Redact messages in the channel.
\item List messages in the channel.
}
}
\section{Request syntax}{
\preformatted{svc$create_channel_moderator(
  ChannelArn = "string",
  ChannelModeratorArn = "string"
)
}
}

\keyword{internal}
