% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_meeting_dial_out}
\alias{chime_create_meeting_dial_out}
\title{Uses the join token and call metadata in a meeting request (From number,
To number, and so forth) to initiate an outbound call to a public
switched telephone network (PSTN) and joins them into Chime meeting}
\usage{
chime_create_meeting_dial_out(MeetingId, FromPhoneNumber, ToPhoneNumber,
  JoinToken)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.

Type: String

Pattern: [a-fA-F0-9]\{8\}(?:-[a-fA-F0-9]\{4\})\{3\}-[a-fA-F0-9]\{12\}

Required: No}

\item{FromPhoneNumber}{[required] Phone number used as the caller ID when the remote party receives a
call.}

\item{ToPhoneNumber}{[required] Phone number called when inviting someone to a meeting.}

\item{JoinToken}{[required] Token used by the Amazon Chime SDK attendee. Call the CreateAttendee API
to get a join token.}
}
\description{
Uses the join token and call metadata in a meeting request (From number,
To number, and so forth) to initiate an outbound call to a public
switched telephone network (PSTN) and joins them into Chime meeting.
Also ensures that the From number belongs to the customer.

To play welcome audio or implement an interactive voice response (IVR),
use the
\code{\link[=chime_create_sip_media_application_call]{create_sip_media_application_call}}
API with the corresponding SIP media application ID.
}
\section{Request syntax}{
\preformatted{svc$create_meeting_dial_out(
  MeetingId = "string",
  FromPhoneNumber = "string",
  ToPhoneNumber = "string",
  JoinToken = "string"
)
}
}

\keyword{internal}
