% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_room}
\alias{chime_create_room}
\title{Creates a chat room for the specified Amazon Chime account}
\usage{
chime_create_room(AccountId, Name, ClientRequestToken)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{Name}{[required] The room name.}

\item{ClientRequestToken}{The idempotency token for the request.}
}
\description{
Creates a chat room for the specified Amazon Chime account.
}
\section{Request syntax}{
\preformatted{svc$create_room(
  AccountId = "string",
  Name = "string",
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
