% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_room_membership}
\alias{chime_create_room_membership}
\title{Adds a member to a chat room}
\usage{
chime_create_room_membership(AccountId, RoomId, MemberId, Role)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RoomId}{[required] The room ID.}

\item{MemberId}{[required] The Amazon Chime member ID (user ID or bot ID).}

\item{Role}{The role of the member.}
}
\description{
Adds a member to a chat room. A member can be either a user or a bot.
The member role designates whether the member is a chat room
administrator or a general chat room member.
}
\section{Request syntax}{
\preformatted{svc$create_room_membership(
  AccountId = "string",
  RoomId = "string",
  MemberId = "string",
  Role = "Administrator"|"Member"
)
}
}

\keyword{internal}
