% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_user}
\alias{chime_create_user}
\title{Creates a user under the specified Amazon Chime account}
\usage{
chime_create_user(AccountId, Username, Email, UserType)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{Username}{The user name.}

\item{Email}{The user\'s email address.}

\item{UserType}{The user type.}
}
\description{
Creates a user under the specified Amazon Chime account.
}
\section{Request syntax}{
\preformatted{svc$create_user(
  AccountId = "string",
  Username = "string",
  Email = "string",
  UserType = "PrivateUser"|"SharedDevice"
)
}
}

\keyword{internal}
