% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_account}
\alias{chime_delete_account}
\title{Deletes the specified Amazon Chime account}
\usage{
chime_delete_account(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}
}
\description{
Deletes the specified Amazon Chime account. You must suspend all users
before deleting a \code{Team} account. You can use the BatchSuspendUser
action to do so.

For \code{EnterpriseLWA} and \code{EnterpriseAD} accounts, you must release the
claimed domains for your Amazon Chime account before deletion. As soon
as you release the domain, all users under that account are suspended.

Deleted accounts appear in your \code{Disabled} accounts list for 90 days. To
restore a deleted account from your \code{Disabled} accounts list, you must
contact AWS Support.

After 90 days, deleted accounts are permanently removed from your
\code{Disabled} accounts list.
}
\section{Request syntax}{
\preformatted{svc$delete_account(
  AccountId = "string"
)
}
}

\keyword{internal}
