% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_attendee}
\alias{chime_delete_attendee}
\title{Deletes an attendee from the specified Amazon Chime SDK meeting and
deletes their JoinToken}
\usage{
chime_delete_attendee(MeetingId, AttendeeId)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{AttendeeId}{[required] The Amazon Chime SDK attendee ID.}
}
\description{
Deletes an attendee from the specified Amazon Chime SDK meeting and
deletes their \code{JoinToken}. Attendees are automatically deleted when a
Amazon Chime SDK meeting is deleted. For more information about the
Amazon Chime SDK, see \href{https://docs.aws.amazon.com/chime/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK} in
the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$delete_attendee(
  MeetingId = "string",
  AttendeeId = "string"
)
}
}

\keyword{internal}
