% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_channel}
\alias{chime_delete_channel}
\title{Immediately makes a channel and its memberships inaccessible and marks
them for deletion}
\usage{
chime_delete_channel(ChannelArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel being deleted.}
}
\description{
Immediately makes a channel and its memberships inaccessible and marks
them for deletion. This is an irreversible process.
}
\section{Request syntax}{
\preformatted{svc$delete_channel(
  ChannelArn = "string"
)
}
}

\keyword{internal}
