% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_room_membership}
\alias{chime_delete_room_membership}
\title{Removes a member from a chat room}
\usage{
chime_delete_room_membership(AccountId, RoomId, MemberId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RoomId}{[required] The room ID.}

\item{MemberId}{[required] The member ID (user ID or bot ID).}
}
\description{
Removes a member from a chat room.
}
\section{Request syntax}{
\preformatted{svc$delete_room_membership(
  AccountId = "string",
  RoomId = "string",
  MemberId = "string"
)
}
}

\keyword{internal}
