% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_app_instance_admin}
\alias{chime_describe_app_instance_admin}
\title{Returns the full details of an AppInstanceAdmin}
\usage{
chime_describe_app_instance_admin(AppInstanceAdminArn, AppInstanceArn)
}
\arguments{
\item{AppInstanceAdminArn}{[required] The ARN of the app instance administrator.}

\item{AppInstanceArn}{[required] The ARN of the app instance.}
}
\description{
Returns the full details of an \code{AppInstanceAdmin}.
}
\section{Request syntax}{
\preformatted{svc$describe_app_instance_admin(
  AppInstanceAdminArn = "string",
  AppInstanceArn = "string"
)
}
}

\keyword{internal}
