% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_app_instance_user}
\alias{chime_describe_app_instance_user}
\title{Returns the full details of an AppInstanceUser}
\usage{
chime_describe_app_instance_user(AppInstanceUserArn)
}
\arguments{
\item{AppInstanceUserArn}{[required] The ARN of the app instance user.}
}
\description{
Returns the full details of an \code{AppInstanceUser}.
}
\section{Request syntax}{
\preformatted{svc$describe_app_instance_user(
  AppInstanceUserArn = "string"
)
}
}

\keyword{internal}
