% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel}
\alias{chime_describe_channel}
\title{Returns the full details of a channel in an Amazon Chime app instance}
\usage{
chime_describe_channel(ChannelArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}
}
\description{
Returns the full details of a channel in an Amazon Chime app instance.
}
\section{Request syntax}{
\preformatted{svc$describe_channel(
  ChannelArn = "string"
)
}
}

\keyword{internal}
