% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_disassociate_phone_numbers_from_voice_connector_group}
\alias{chime_disassociate_phone_numbers_from_voice_connector_group}
\title{Disassociates the specified phone numbers from the specified Amazon
Chime Voice Connector group}
\usage{
chime_disassociate_phone_numbers_from_voice_connector_group(
  VoiceConnectorGroupId, E164PhoneNumbers)
}
\arguments{
\item{VoiceConnectorGroupId}{[required] The Amazon Chime Voice Connector group ID.}

\item{E164PhoneNumbers}{List of phone numbers, in E.164 format.}
}
\description{
Disassociates the specified phone numbers from the specified Amazon
Chime Voice Connector group.
}
\section{Request syntax}{
\preformatted{svc$disassociate_phone_numbers_from_voice_connector_group(
  VoiceConnectorGroupId = "string",
  E164PhoneNumbers = list(
    "string"
  )
)
}
}

\keyword{internal}
