% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_messaging_session_endpoint}
\alias{chime_get_messaging_session_endpoint}
\title{The endpoint for the messaging session}
\usage{
chime_get_messaging_session_endpoint()
}
\description{
The endpoint for the messaging session.
}
\section{Request syntax}{
\preformatted{svc$get_messaging_session_endpoint()
}
}

\keyword{internal}
