% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_retention_settings}
\alias{chime_get_retention_settings}
\title{Gets the retention settings for the specified Amazon Chime Enterprise
account}
\usage{
chime_get_retention_settings(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}
}
\description{
Gets the retention settings for the specified Amazon Chime Enterprise
account. For more information about retention settings, see \href{https://docs.aws.amazon.com/chime/latest/ag/chat-retention.html}{Managing Chat Retention Policies}
in the \emph{Amazon Chime Administration Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_retention_settings(
  AccountId = "string"
)
}
}

\keyword{internal}
