% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_user}
\alias{chime_get_user}
\title{Retrieves details for the specified user ID, such as primary email
address, license type, and personal meeting PIN}
\usage{
chime_get_user(AccountId, UserId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}
}
\description{
Retrieves details for the specified user ID, such as primary email
address, license type, and personal meeting PIN.
}
\details{
To retrieve user details with an email address instead of a user ID, use
the ListUsers action, and then filter by email address.
}
\section{Request syntax}{
\preformatted{chime$get_user(
  AccountId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
