% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector_termination}
\alias{chime_get_voice_connector_termination}
\title{Retrieves termination setting details for the specified Amazon Chime
Voice Connector}
\usage{
chime_get_voice_connector_termination(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\description{
Retrieves termination setting details for the specified Amazon Chime
Voice Connector.
}
\section{Request syntax}{
\preformatted{svc$get_voice_connector_termination(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
