% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_app_instance_users}
\alias{chime_list_app_instance_users}
\title{List all AppInstanceUsers created under a single app instance}
\usage{
chime_list_app_instance_users(AppInstanceArn, MaxResults, NextToken)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the app instance.}

\item{MaxResults}{The maximum number of requests that you want returned.}

\item{NextToken}{The token passed by previous API calls until all requested users are
returned.}
}
\description{
List all \code{AppInstanceUsers} created under a single app instance.
}
\section{Request syntax}{
\preformatted{svc$list_app_instance_users(
  AppInstanceArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
