% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_attendees}
\alias{chime_list_attendees}
\title{Lists the attendees for the specified Amazon Chime SDK meeting}
\usage{
chime_list_attendees(MeetingId, NextToken, MaxResults)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{NextToken}{The token to use to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Lists the attendees for the specified Amazon Chime SDK meeting. For more
information about the Amazon Chime SDK, see \href{https://docs.aws.amazon.com/chime/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK} in
the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$list_attendees(
  MeetingId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
