% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_channel_memberships}
\alias{chime_list_channel_memberships}
\title{Lists all channel memberships in a channel}
\usage{
chime_list_channel_memberships(ChannelArn, Type, MaxResults, NextToken)
}
\arguments{
\item{ChannelArn}{[required] The maximum number of channel memberships that you want returned.}

\item{Type}{The membership type of a user, \code{DEFAULT} or \code{HIDDEN}. Default members
are always returned as part of
\code{\link[=chime_list_channel_memberships]{list_channel_memberships}}. Hidden
members are only returned if the type filter in
\code{\link[=chime_list_channel_memberships]{list_channel_memberships}} equals
\code{HIDDEN}. Otherwise hidden members are not returned.}

\item{MaxResults}{The maximum number of channel memberships that you want returned.}

\item{NextToken}{The token passed by previous API calls until all requested channel
memberships are returned..}
}
\description{
Lists all channel memberships in a channel.
}
\section{Request syntax}{
\preformatted{svc$list_channel_memberships(
  ChannelArn = "string",
  Type = "DEFAULT"|"HIDDEN",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
