% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_proxy_sessions}
\alias{chime_list_proxy_sessions}
\title{Lists the proxy sessions for the specified Amazon Chime Voice Connector}
\usage{
chime_list_proxy_sessions(VoiceConnectorId, Status, NextToken,
  MaxResults)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime voice connector ID.}

\item{Status}{The proxy session status.}

\item{NextToken}{The token to use to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Lists the proxy sessions for the specified Amazon Chime Voice Connector.
}
\section{Request syntax}{
\preformatted{svc$list_proxy_sessions(
  VoiceConnectorId = "string",
  Status = "Open"|"InProgress"|"Closed",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
