% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_sip_media_applications}
\alias{chime_list_sip_media_applications}
\title{Lists the SIP media applications under the administrator's AWS account}
\usage{
chime_list_sip_media_applications(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in a single call. Defaults to
100.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\description{
Lists the SIP media applications under the administrator's AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_sip_media_applications(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
