% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_sip_media_application_logging_configuration}
\alias{chime_put_sip_media_application_logging_configuration}
\title{Updates the logging configuration for the specified SIP media
application}
\usage{
chime_put_sip_media_application_logging_configuration(
  SipMediaApplicationId, SipMediaApplicationLoggingConfiguration)
}
\arguments{
\item{SipMediaApplicationId}{[required] The ID of the specified SIP media application}

\item{SipMediaApplicationLoggingConfiguration}{The actual logging configuration.}
}
\description{
Updates the logging configuration for the specified SIP media
application.
}
\section{Request syntax}{
\preformatted{svc$put_sip_media_application_logging_configuration(
  SipMediaApplicationId = "string",
  SipMediaApplicationLoggingConfiguration = list(
    EnableSipMediaApplicationMessageLogs = TRUE|FALSE
  )
)
}
}

\keyword{internal}
