% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_voice_connector_termination}
\alias{chime_put_voice_connector_termination}
\title{Adds termination settings for the specified Amazon Chime Voice Connector}
\usage{
chime_put_voice_connector_termination(VoiceConnectorId, Termination)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}

\item{Termination}{[required] The termination setting details to add.}
}
\description{
Adds termination settings for the specified Amazon Chime Voice
Connector.

If emergency calling is configured for the Amazon Chime Voice Connector,
it must be deleted prior to turning off termination settings.
}
\section{Request syntax}{
\preformatted{svc$put_voice_connector_termination(
  VoiceConnectorId = "string",
  Termination = list(
    CpsLimit = 123,
    DefaultPhoneNumber = "string",
    CallingRegions = list(
      "string"
    ),
    CidrAllowedList = list(
      "string"
    ),
    Disabled = TRUE|FALSE
  )
)
}
}

\keyword{internal}
