% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_regenerate_security_token}
\alias{chime_regenerate_security_token}
\title{Regenerates the security token for a bot}
\usage{
chime_regenerate_security_token(AccountId, BotId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{BotId}{[required] The bot ID.}
}
\description{
Regenerates the security token for a bot.
}
\section{Request syntax}{
\preformatted{svc$regenerate_security_token(
  AccountId = "string",
  BotId = "string"
)
}
}

\keyword{internal}
