% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_restore_phone_number}
\alias{chime_restore_phone_number}
\title{Moves a phone number from the Deletion queue back into the phone number
Inventory}
\usage{
chime_restore_phone_number(PhoneNumberId)
}
\arguments{
\item{PhoneNumberId}{[required] The phone number.}
}
\description{
Moves a phone number from the \strong{Deletion queue} back into the phone
number \strong{Inventory}.
}
\section{Request syntax}{
\preformatted{svc$restore_phone_number(
  PhoneNumberId = "string"
)
}
}

\keyword{internal}
