% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_search_available_phone_numbers}
\alias{chime_search_available_phone_numbers}
\title{Searches phone numbers that can be ordered}
\usage{
chime_search_available_phone_numbers(AreaCode, City, Country, State,
  TollFreePrefix, MaxResults, NextToken)
}
\arguments{
\item{AreaCode}{The area code used to filter results.}

\item{City}{The city used to filter results.}

\item{Country}{The country used to filter results.}

\item{State}{The state used to filter results.}

\item{TollFreePrefix}{The toll-free prefix that you use to filter results.}

\item{MaxResults}{The maximum number of results to return in a single call.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\description{
Searches phone numbers that can be ordered.
}
\section{Request syntax}{
\preformatted{svc$search_available_phone_numbers(
  AreaCode = "string",
  City = "string",
  Country = "string",
  State = "string",
  TollFreePrefix = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
