% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_tag_attendee}
\alias{chime_tag_attendee}
\title{Applies the specified tags to the specified Amazon Chime SDK attendee}
\usage{
chime_tag_attendee(MeetingId, AttendeeId, Tags)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{AttendeeId}{[required] The Amazon Chime SDK attendee ID.}

\item{Tags}{[required] The tag key-value pairs.}
}
\description{
Applies the specified tags to the specified Amazon Chime SDK attendee.
}
\section{Request syntax}{
\preformatted{svc$tag_attendee(
  MeetingId = "string",
  AttendeeId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
