% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_untag_meeting}
\alias{chime_untag_meeting}
\title{Untags the specified tags from the specified Amazon Chime SDK meeting}
\usage{
chime_untag_meeting(MeetingId, TagKeys)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{TagKeys}{[required] The tag keys.}
}
\description{
Untags the specified tags from the specified Amazon Chime SDK meeting.
}
\section{Request syntax}{
\preformatted{svc$untag_meeting(
  MeetingId = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
