% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_stack_instance}
\alias{cloudformation_describe_stack_instance}
\title{Returns the stack instance that's associated with the specified stack
set, AWS account, and Region}
\usage{
cloudformation_describe_stack_instance(StackSetName,
  StackInstanceAccount, StackInstanceRegion)
}
\arguments{
\item{StackSetName}{[required] The name or the unique stack ID of the stack set that you want to get
stack instance information for.}

\item{StackInstanceAccount}{[required] The ID of an AWS account that's associated with this stack instance.}

\item{StackInstanceRegion}{[required] The name of a Region that's associated with this stack instance.}
}
\description{
Returns the stack instance that's associated with the specified stack
set, AWS account, and Region.

For a list of stack instances that are associated with a specific stack
set, use \code{\link[=cloudformation_list_stack_instances]{list_stack_instances}}.
}
\section{Request syntax}{
\preformatted{svc$describe_stack_instance(
  StackSetName = "string",
  StackInstanceAccount = "string",
  StackInstanceRegion = "string"
)
}
}

\keyword{internal}
