% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_stack_resource}
\alias{cloudformation_describe_stack_resource}
\title{Returns a description of the specified resource in the specified stack}
\usage{
cloudformation_describe_stack_resource(StackName, LogicalResourceId)
}
\arguments{
\item{StackName}{[required] The name or the unique stack ID that is associated with the stack, which
are not always interchangeable:
\itemize{
\item Running stacks: You can specify either the stack\'s name or its
unique stack ID.
\item Deleted stacks: You must specify the unique stack ID.
}

Default: There is no default value.}

\item{LogicalResourceId}{[required] The logical name of the resource as specified in the template.

Default: There is no default value.}
}
\description{
Returns a description of the specified resource in the specified stack.
}
\details{
For deleted stacks, DescribeStackResource returns resource information
for up to 90 days after the stack has been deleted.
}
\section{Request syntax}{
\preformatted{svc$describe_stack_resource(
  StackName = "string",
  LogicalResourceId = "string"
)
}
}

\keyword{internal}
