% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_stack_set}
\alias{cloudformation_describe_stack_set}
\title{Returns the description of the specified stack set}
\usage{
cloudformation_describe_stack_set(StackSetName)
}
\arguments{
\item{StackSetName}{[required] The name or unique ID of the stack set whose description you want.}
}
\description{
Returns the description of the specified stack set.
}
\section{Request syntax}{
\preformatted{cloudformation$describe_stack_set(
  StackSetName = "string"
)
}
}

\keyword{internal}
