% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_list_stack_resources}
\alias{cloudformation_list_stack_resources}
\title{Returns descriptions of all resources of the specified stack}
\usage{
cloudformation_list_stack_resources(StackName, NextToken)
}
\arguments{
\item{StackName}{[required] The name or the unique stack ID that is associated with the stack, which
are not always interchangeable:
\itemize{
\item Running stacks: You can specify either the stack\'s name or its
unique stack ID.
\item Deleted stacks: You must specify the unique stack ID.
}

Default: There is no default value.}

\item{NextToken}{A string that identifies the next page of stack resources that you want
to retrieve.}
}
\description{
Returns descriptions of all resources of the specified stack.
}
\details{
For deleted stacks, ListStackResources returns resource information for
up to 90 days after the stack has been deleted.
}
\section{Request syntax}{
\preformatted{svc$list_stack_resources(
  StackName = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
