% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_invalidation}
\alias{cloudfront_create_invalidation}
\title{Create a new invalidation}
\usage{
cloudfront_create_invalidation(DistributionId, InvalidationBatch)
}
\arguments{
\item{DistributionId}{[required] The distribution's id.}

\item{InvalidationBatch}{[required] The batch information for the invalidation.}
}
\description{
Create a new invalidation.
}
\section{Request syntax}{
\preformatted{svc$create_invalidation(
  DistributionId = "string",
  InvalidationBatch = list(
    Paths = list(
      Quantity = 123,
      Items = list(
        "string"
      )
    ),
    CallerReference = "string"
  )
)
}
}

\keyword{internal}
