% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_cache_policy}
\alias{cloudfront_get_cache_policy}
\title{Gets a cache policy, including the following metadata: - The policy’s
identifier}
\usage{
cloudfront_get_cache_policy(Id)
}
\arguments{
\item{Id}{[required] The unique identifier for the cache policy. If the cache policy is
attached to a distribution’s cache behavior, you can get the policy’s
identifier using \code{ListDistributions} or \code{GetDistribution}. If the cache
policy is not attached to a cache behavior, you can get the identifier
using \code{ListCachePolicies}.}
}
\description{
Gets a cache policy, including the following metadata:
}
\details{
\itemize{
\item The policy’s identifier.
\item The date and time when the policy was last modified.
}

To get a cache policy, you must provide the policy’s identifier. If the
cache policy is attached to a distribution’s cache behavior, you can get
the policy’s identifier using \code{ListDistributions} or \code{GetDistribution}.
If the cache policy is not attached to a cache behavior, you can get the
identifier using \code{ListCachePolicies}.
}
\section{Request syntax}{
\preformatted{svc$get_cache_policy(
  Id = "string"
)
}
}

\keyword{internal}
