% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_cache_policy_config}
\alias{cloudfront_get_cache_policy_config}
\title{Gets a cache policy configuration}
\usage{
cloudfront_get_cache_policy_config(Id)
}
\arguments{
\item{Id}{[required] The unique identifier for the cache policy. If the cache policy is
attached to a distribution’s cache behavior, you can get the policy’s
identifier using \code{ListDistributions} or \code{GetDistribution}. If the cache
policy is not attached to a cache behavior, you can get the identifier
using \code{ListCachePolicies}.}
}
\description{
Gets a cache policy configuration.

To get a cache policy configuration, you must provide the policy’s
identifier. If the cache policy is attached to a distribution’s cache
behavior, you can get the policy’s identifier using \code{ListDistributions}
or \code{GetDistribution}. If the cache policy is not attached to a cache
behavior, you can get the identifier using \code{ListCachePolicies}.
}
\section{Request syntax}{
\preformatted{svc$get_cache_policy_config(
  Id = "string"
)
}
}

\keyword{internal}
