% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_key_group_config}
\alias{cloudfront_get_key_group_config}
\title{Gets a key group configuration}
\usage{
cloudfront_get_key_group_config(Id)
}
\arguments{
\item{Id}{[required] The identifier of the key group whose configuration you are getting. To
get the identifier, use \code{ListKeyGroups}.}
}
\description{
Gets a key group configuration.
}
\details{
To get a key group configuration, you must provide the key group’s
identifier. If the key group is referenced in a distribution’s cache
behavior, you can get the key group’s identifier using
\code{ListDistributions} or \code{GetDistribution}. If the key group is not
referenced in a cache behavior, you can get the identifier using
\code{ListKeyGroups}.
}
\section{Request syntax}{
\preformatted{svc$get_key_group_config(
  Id = "string"
)
}
}

\keyword{internal}
