% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_origin_request_policy_config}
\alias{cloudfront_get_origin_request_policy_config}
\title{Gets an origin request policy configuration}
\usage{
cloudfront_get_origin_request_policy_config(Id)
}
\arguments{
\item{Id}{[required] The unique identifier for the origin request policy. If the origin
request policy is attached to a distribution’s cache behavior, you can
get the policy’s identifier using \code{ListDistributions} or
\code{GetDistribution}. If the origin request policy is not attached to a
cache behavior, you can get the identifier using
\code{ListOriginRequestPolicies}.}
}
\description{
Gets an origin request policy configuration.
}
\details{
To get an origin request policy configuration, you must provide the
policy’s identifier. If the origin request policy is attached to a
distribution’s cache behavior, you can get the policy’s identifier using
\code{ListDistributions} or \code{GetDistribution}. If the origin request policy
is not attached to a cache behavior, you can get the identifier using
\code{ListOriginRequestPolicies}.
}
\section{Request syntax}{
\preformatted{svc$get_origin_request_policy_config(
  Id = "string"
)
}
}

\keyword{internal}
