% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_public_key}
\alias{cloudfront_get_public_key}
\title{Gets a public key}
\usage{
cloudfront_get_public_key(Id)
}
\arguments{
\item{Id}{[required] The identifier of the public key you are getting.}
}
\description{
Gets a public key.
}
\section{Request syntax}{
\preformatted{svc$get_public_key(
  Id = "string"
)
}
}

\keyword{internal}
