% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_streaming_distributions}
\alias{cloudfront_list_streaming_distributions}
\title{List streaming distributions}
\usage{
cloudfront_list_streaming_distributions(Marker, MaxItems)
}
\arguments{
\item{Marker}{The value that you provided for the \code{Marker} request parameter.}

\item{MaxItems}{The value that you provided for the \code{MaxItems} request parameter.}
}
\description{
List streaming distributions.
}
\section{Request syntax}{
\preformatted{svc$list_streaming_distributions(
  Marker = "string",
  MaxItems = "string"
)
}
}

\keyword{internal}
