% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_tag_resource}
\alias{cloudfront_tag_resource}
\title{Add tags to a CloudFront resource}
\usage{
cloudfront_tag_resource(Resource, Tags)
}
\arguments{
\item{Resource}{[required] An ARN of a CloudFront resource.}

\item{Tags}{[required] A complex type that contains zero or more \code{Tag} elements.}
}
\description{
Add tags to a CloudFront resource.
}
\section{Request syntax}{
\preformatted{cloudfront$tag_resource(
  Resource = "string",
  Tags = list(
    Items = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}
}

\keyword{internal}
