% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_list_luna_clients}
\alias{cloudhsm_list_luna_clients}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_list_luna_clients(NextToken)
}
\arguments{
\item{NextToken}{The \code{NextToken} value from a previous call to \code{ListLunaClients}. Pass
null if this is the first call.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more
information, see \href{http://aws.amazon.com/cloudhsm/faqs-classic/}{AWS CloudHSM Classic FAQs}, the \href{http://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide}, and the
\href{http://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.
}
\details{
\strong{For information about the current version of AWS CloudHSM}, see \href{http://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{http://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the
\href{http://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Lists all of the clients.

This operation supports pagination with the use of the \code{NextToken}
member. If more results are available, the \code{NextToken} member of the
response contains a token that you pass in the next call to
\code{ListLunaClients} to retrieve the next set of items.
}
\section{Request syntax}{
\preformatted{svc$list_luna_clients(
  NextToken = "string"
)
}
}

\keyword{internal}
