% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_delete_backup}
\alias{cloudhsmv2_delete_backup}
\title{Deletes a specified AWS CloudHSM backup}
\usage{
cloudhsmv2_delete_backup(BackupId)
}
\arguments{
\item{BackupId}{[required] The ID of the backup to be deleted. To find the ID of a backup, use the
DescribeBackups operation.}
}
\description{
Deletes a specified AWS CloudHSM backup. A backup can be restored up to
7 days after the DeleteBackup request. For more information on restoring
a backup, see RestoreBackup
}
\section{Request syntax}{
\preformatted{cloudhsmv2$delete_backup(
  BackupId = "string"
)
}
}

\keyword{internal}
