% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_delete_cluster}
\alias{cloudhsmv2_delete_cluster}
\title{Deletes the specified AWS CloudHSM cluster}
\usage{
cloudhsmv2_delete_cluster(ClusterId)
}
\arguments{
\item{ClusterId}{[required] The identifier (ID) of the cluster that you are deleting. To find the
cluster ID, use DescribeClusters.}
}
\description{
Deletes the specified AWS CloudHSM cluster. Before you can delete a
cluster, you must delete all HSMs in the cluster. To see if the cluster
contains any HSMs, use DescribeClusters. To delete an HSM, use
DeleteHsm.
}
\section{Request syntax}{
\preformatted{cloudhsmv2$delete_cluster(
  ClusterId = "string"
)
}
}

\keyword{internal}
