% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_describe_scaling_parameters}
\alias{cloudsearch_describe_scaling_parameters}
\title{Gets the scaling parameters configured for a domain}
\usage{
cloudsearch_describe_scaling_parameters(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\description{
Gets the scaling parameters configured for a domain. A domain\'s scaling
parameters specify the desired search instance type and replication
count. For more information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-scaling-options.html}{Configuring Scaling Options}
in the \emph{Amazon CloudSearch Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$describe_scaling_parameters(
  DomainName = "string"
)
}
}

\keyword{internal}
