% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_delete_dashboards}
\alias{cloudwatch_delete_dashboards}
\title{Deletes all dashboards that you specify}
\usage{
cloudwatch_delete_dashboards(DashboardNames)
}
\arguments{
\item{DashboardNames}{[required] The dashboards to be deleted. This parameter is required.}
}
\description{
Deletes all dashboards that you specify. You may specify up to 100
dashboards to delete. If there is an error during this call, no
dashboards are deleted.
}
\section{Request syntax}{
\preformatted{svc$delete_dashboards(
  DashboardNames = list(
    "string"
  )
)
}
}

\keyword{internal}
