% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_delete_insight_rules}
\alias{cloudwatch_delete_insight_rules}
\title{Permanently deletes the specified Contributor Insights rules}
\usage{
cloudwatch_delete_insight_rules(RuleNames)
}
\arguments{
\item{RuleNames}{[required] An array of the rule names to delete. If you need to find out the names
of your rules, use
\href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeInsightRules.html}{DescribeInsightRules}.}
}
\description{
Permanently deletes the specified Contributor Insights rules.
}
\details{
If you create a rule, delete it, and then re-create it with the same
name, historical data from the first time the rule was created may or
may not be available.
}
\section{Request syntax}{
\preformatted{svc$delete_insight_rules(
  RuleNames = list(
    "string"
  )
)
}
}

\keyword{internal}
