% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_describe_insight_rules}
\alias{cloudwatch_describe_insight_rules}
\title{Returns a list of all the Contributor Insights rules in your account}
\usage{
cloudwatch_describe_insight_rules(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{Reserved for future use.}

\item{MaxResults}{This parameter is not currently used. Reserved for future use. If it is
used in the future, the maximum value may be different.}
}
\description{
Returns a list of all the Contributor Insights rules in your account.
All rules in your account are returned with a single operation.
}
\details{
For more information about Contributor Insights, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/ContributorInsights.html}{Using Contributor Insights to Analyze High-Cardinality Data}.
}
\section{Request syntax}{
\preformatted{svc$describe_insight_rules(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
