% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_enable_insight_rules}
\alias{cloudwatch_enable_insight_rules}
\title{Enables the specified Contributor Insights rules}
\usage{
cloudwatch_enable_insight_rules(RuleNames)
}
\arguments{
\item{RuleNames}{[required] An array of the rule names to enable. If you need to find out the names
of your rules, use
\code{\link[=cloudwatch_describe_insight_rules]{describe_insight_rules}}.}
}
\description{
Enables the specified Contributor Insights rules. When rules are
enabled, they immediately begin analyzing log data.
}
\section{Request syntax}{
\preformatted{svc$enable_insight_rules(
  RuleNames = list(
    "string"
  )
)
}
}

\keyword{internal}
