% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_describe_rule}
\alias{cloudwatchevents_describe_rule}
\title{Describes the specified rule}
\usage{
cloudwatchevents_describe_rule(Name)
}
\arguments{
\item{Name}{[required] The name of the rule.}
}
\description{
Describes the specified rule.
}
\details{
DescribeRule does not list the targets of a rule. To see the targets
associated with a rule, use ListTargetsByRule.
}
\section{Request syntax}{
\preformatted{cloudwatchevents$describe_rule(
  Name = "string"
)
}
}

\keyword{internal}
