% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_destination}
\alias{cloudwatchlogs_delete_destination}
\title{Deletes the specified destination, and eventually disables all the
subscription filters that publish to it}
\usage{
cloudwatchlogs_delete_destination(destinationName)
}
\arguments{
\item{destinationName}{[required] The name of the destination.}
}
\description{
Deletes the specified destination, and eventually disables all the
subscription filters that publish to it. This operation does not delete
the physical resource encapsulated by the destination.
}
\section{Request syntax}{
\preformatted{svc$delete_destination(
  destinationName = "string"
)
}
}

\keyword{internal}
