% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_resource_policy}
\alias{cloudwatchlogs_delete_resource_policy}
\title{Deletes a resource policy from this account}
\usage{
cloudwatchlogs_delete_resource_policy(policyName)
}
\arguments{
\item{policyName}{The name of the policy to be revoked. This parameter is required.}
}
\description{
Deletes a resource policy from this account. This revokes the access of
the identities in that policy to put log events to this account.
}
\section{Request syntax}{
\preformatted{cloudwatchlogs$delete_resource_policy(
  policyName = "string"
)
}
}

\keyword{internal}
