% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_query_definitions}
\alias{cloudwatchlogs_describe_query_definitions}
\title{Describe query definitions}
\usage{
cloudwatchlogs_describe_query_definitions(queryDefinitionNamePrefix,
  maxResults, nextToken)
}
\arguments{
\item{queryDefinitionNamePrefix}{}

\item{maxResults}{}

\item{nextToken}{}
}
\description{
Describe query definitions
}
\section{Request syntax}{
\preformatted{svc$describe_query_definitions(
  queryDefinitionNamePrefix = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
