% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_stop_query}
\alias{cloudwatchlogs_stop_query}
\title{Stops a CloudWatch Logs Insights query that is in progress}
\usage{
cloudwatchlogs_stop_query(queryId)
}
\arguments{
\item{queryId}{[required] The ID number of the query to stop. To find this ID number, use
\code{DescribeQueries}.}
}
\description{
Stops a CloudWatch Logs Insights query that is in progress. If the query
has already ended, the operation returns an error indicating that the
specified query is not running.
}
\section{Request syntax}{
\preformatted{svc$stop_query(
  queryId = "string"
)
}
}

\keyword{internal}
