% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_delete_repository_permissions_policy}
\alias{codeartifact_delete_repository_permissions_policy}
\title{Deletes the resource policy that is set on a repository}
\usage{
codeartifact_delete_repository_permissions_policy(domain, domainOwner,
  repository, policyRevision)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the repository associated with the
resource policy to be deleted.}

\item{domainOwner}{The 12-digit account number of the AWS account that owns the domain. It
does not include dashes or spaces.}

\item{repository}{[required] The name of the repository that is associated with the resource policy
to be deleted}

\item{policyRevision}{The revision of the repository's resource policy to be deleted. This
revision is used for optimistic locking, which prevents others from
accidentally overwriting your changes to the repository's resource
policy.}
}
\description{
Deletes the resource policy that is set on a repository. After a
resource policy is deleted, the permissions allowed and denied by the
deleted policy are removed. The effect of deleting a resource policy
might not be immediate.
}
\details{
Use \code{DeleteRepositoryPermissionsPolicy} with caution. After a policy is
deleted, AWS users, roles, and accounts lose permissions to perform the
repository actions granted by the deleted policy.
}
\section{Request syntax}{
\preformatted{svc$delete_repository_permissions_policy(
  domain = "string",
  domainOwner = "string",
  repository = "string",
  policyRevision = "string"
)
}
}

\keyword{internal}
