% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_describe_domain}
\alias{codeartifact_describe_domain}
\title{Returns a DomainDescription object that contains information about the
requested domain}
\usage{
codeartifact_describe_domain(domain, domainOwner)
}
\arguments{
\item{domain}{[required] A string that specifies the name of the requested domain.}

\item{domainOwner}{The 12-digit account number of the AWS account that owns the domain. It
does not include dashes or spaces.}
}
\description{
Returns a
\href{https://docs.aws.amazon.com/codeartifact/latest/APIReference/API_DomainDescription.html}{\code{DomainDescription}}
object that contains information about the requested domain.
}
\section{Request syntax}{
\preformatted{svc$describe_domain(
  domain = "string",
  domainOwner = "string"
)
}
}

\keyword{internal}
