% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_describe_package_version}
\alias{codeartifact_describe_package_version}
\title{Returns a PackageVersionDescription object that contains information
about the requested package version}
\usage{
codeartifact_describe_package_version(domain, domainOwner, repository,
  format, namespace, package, packageVersion)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the repository that contains the
package version.}

\item{domainOwner}{The 12-digit account number of the AWS account that owns the domain. It
does not include dashes or spaces.}

\item{repository}{[required] The name of the repository that contains the package version.}

\item{format}{[required] A format that specifies the type of the requested package version. The
valid values are:
\itemize{
\item \code{npm}
\item \code{pypi}
\item \code{maven}
\item \code{nuget}
}}

\item{namespace}{The namespace of the package. The package component that specifies its
namespace depends on its type. For example:
\itemize{
\item The namespace of a Maven package is its \code{groupId}.
\item The namespace of an npm package is its \code{scope}.
\item A Python package does not contain a corresponding component, so
Python packages do not have a namespace.
\item A NuGet package does not contain a corresponding component, so NuGet
packages do not have a namespace.
}}

\item{package}{[required] The name of the requested package version.}

\item{packageVersion}{[required] A string that contains the package version (for example, \verb{3.5.2}).}
}
\description{
Returns a
\href{https://docs.aws.amazon.com/codeartifact/latest/APIReference/API_PackageVersionDescription.html}{\code{PackageVersionDescription}}
object that contains information about the requested package version.
}
\section{Request syntax}{
\preformatted{svc$describe_package_version(
  domain = "string",
  domainOwner = "string",
  repository = "string",
  format = "npm"|"pypi"|"maven"|"nuget",
  namespace = "string",
  package = "string",
  packageVersion = "string"
)
}
}

\keyword{internal}
