% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_get_domain_permissions_policy}
\alias{codeartifact_get_domain_permissions_policy}
\title{Returns the resource policy attached to the specified domain}
\usage{
codeartifact_get_domain_permissions_policy(domain, domainOwner)
}
\arguments{
\item{domain}{[required] The name of the domain to which the resource policy is attached.}

\item{domainOwner}{The 12-digit account number of the AWS account that owns the domain. It
does not include dashes or spaces.}
}
\description{
Returns the resource policy attached to the specified domain.
}
\details{
The policy is a resource-based policy, not an identity-based policy. For
more information, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access_policies_identity-vs-resource.html}{Identity-based policies and resource-based policies}
in the \emph{AWS Identity and Access Management User Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_domain_permissions_policy(
  domain = "string",
  domainOwner = "string"
)
}
}

\keyword{internal}
