% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_get_repository_permissions_policy}
\alias{codeartifact_get_repository_permissions_policy}
\title{Returns the resource policy that is set on a repository}
\usage{
codeartifact_get_repository_permissions_policy(domain, domainOwner,
  repository)
}
\arguments{
\item{domain}{[required] The name of the domain containing the repository whose associated
resource policy is to be retrieved.}

\item{domainOwner}{The 12-digit account number of the AWS account that owns the domain. It
does not include dashes or spaces.}

\item{repository}{[required] The name of the repository whose associated resource policy is to be
retrieved.}
}
\description{
Returns the resource policy that is set on a repository.
}
\section{Request syntax}{
\preformatted{svc$get_repository_permissions_policy(
  domain = "string",
  domainOwner = "string",
  repository = "string"
)
}
}

\keyword{internal}
