% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_list_domains}
\alias{codeartifact_list_domains}
\title{Returns a list of DomainSummary objects for all domains owned by the AWS
account that makes this call}
\usage{
codeartifact_list_domains(maxResults, nextToken)
}
\arguments{
\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Returns a list of \code{DomainSummary} objects for all domains owned by the
AWS account that makes this call. Each returned \code{DomainSummary} object
contains information about a domain.
}
\section{Request syntax}{
\preformatted{svc$list_domains(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
