% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_list_package_version_assets}
\alias{codeartifact_list_package_version_assets}
\title{Returns a list of AssetSummary objects for assets in a package version}
\usage{
codeartifact_list_package_version_assets(domain, domainOwner,
  repository, format, namespace, package, packageVersion, maxResults,
  nextToken)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the repository associated with the
package version assets.}

\item{domainOwner}{The 12-digit account number of the AWS account that owns the domain. It
does not include dashes or spaces.}

\item{repository}{[required] The name of the repository that contains the package that contains the
returned package version assets.}

\item{format}{[required] The format of the package that contains the returned package version
assets. The valid package types are:
\itemize{
\item \code{npm}: A Node Package Manager (npm) package.
\item \code{pypi}: A Python Package Index (PyPI) package.
\item \code{maven}: A Maven package that contains compiled code in a
distributable format, such as a JAR file.
\item \code{nuget}: A NuGet package.
}}

\item{namespace}{The namespace of the package. The package component that specifies its
namespace depends on its type. For example:
\itemize{
\item The namespace of a Maven package is its \code{groupId}.
\item The namespace of an npm package is its \code{scope}.
\item A Python package does not contain a corresponding component, so
Python packages do not have a namespace.
\item A NuGet package does not contain a corresponding component, so NuGet
packages do not have a namespace.
}}

\item{package}{[required] The name of the package that contains the returned package version
assets.}

\item{packageVersion}{[required] A string that contains the package version (for example, \verb{3.5.2}).}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Returns a list of
\href{https://docs.aws.amazon.com/codeartifact/latest/APIReference/API_AssetSummary.html}{\code{AssetSummary}}
objects for assets in a package version.
}
\section{Request syntax}{
\preformatted{svc$list_package_version_assets(
  domain = "string",
  domainOwner = "string",
  repository = "string",
  format = "npm"|"pypi"|"maven"|"nuget",
  namespace = "string",
  package = "string",
  packageVersion = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
