% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_list_repositories}
\alias{codeartifact_list_repositories}
\title{Returns a list of RepositorySummary objects}
\usage{
codeartifact_list_repositories(repositoryPrefix, maxResults, nextToken)
}
\arguments{
\item{repositoryPrefix}{A prefix used to filter returned repositories. Only repositories with
names that start with \code{repositoryPrefix} are returned.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Returns a list of
\href{https://docs.aws.amazon.com/codeartifact/latest/APIReference/API_RepositorySummary.html}{\code{RepositorySummary}}
objects. Each \code{RepositorySummary} contains information about a
repository in the specified AWS account and that matches the input
parameters.
}
\section{Request syntax}{
\preformatted{svc$list_repositories(
  repositoryPrefix = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
