% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_list_tags_for_resource}
\alias{codeartifact_list_tags_for_resource}
\title{Gets information about AWS tags for a specified Amazon Resource Name
(ARN) in AWS CodeArtifact}
\usage{
codeartifact_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to get tags for.}
}
\description{
Gets information about AWS tags for a specified Amazon Resource Name
(ARN) in AWS CodeArtifact.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
