% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_batch_get_builds}
\alias{codebuild_batch_get_builds}
\title{Gets information about one or more builds}
\usage{
codebuild_batch_get_builds(ids)
}
\arguments{
\item{ids}{[required] The IDs of the builds.}
}
\description{
Gets information about one or more builds.
}
\section{Request syntax}{
\preformatted{svc$batch_get_builds(
  ids = list(
    "string"
  )
)
}
}

\examples{
# The following example gets information about builds with the specified
# build IDs.
\dontrun{svc$batch_get_builds(
  ids = list(
    "codebuild-demo-project:9b0ac37f-d19e-4254-9079-f47e9a389eEX",
    "codebuild-demo-project:b79a46f7-1473-4636-a23f-da9c45c208EX"
  )
)}

}
\keyword{internal}
