% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_batch_get_reports}
\alias{codebuild_batch_get_reports}
\title{Returns an array of reports}
\usage{
codebuild_batch_get_reports(reportArns)
}
\arguments{
\item{reportArns}{[required] An array of ARNs that identify the \code{Report} objects to return.}
}
\description{
Returns an array of reports.
}
\section{Request syntax}{
\preformatted{svc$batch_get_reports(
  reportArns = list(
    "string"
  )
)
}
}

\keyword{internal}
