% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_list_curated_environment_images}
\alias{codebuild_list_curated_environment_images}
\title{Gets information about Docker images that are managed by AWS CodeBuild}
\usage{
codebuild_list_curated_environment_images()
}
\description{
Gets information about Docker images that are managed by AWS CodeBuild.
}
\section{Request syntax}{
\preformatted{svc$list_curated_environment_images()
}
}

\keyword{internal}
