% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_stop_build_batch}
\alias{codebuild_stop_build_batch}
\title{Stops a running batch build}
\usage{
codebuild_stop_build_batch(id)
}
\arguments{
\item{id}{[required] The identifier of the batch build to stop.}
}
\description{
Stops a running batch build.
}
\section{Request syntax}{
\preformatted{svc$stop_build_batch(
  id = "string"
)
}
}

\keyword{internal}
