% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_describe_pull_request_events}
\alias{codecommit_describe_pull_request_events}
\title{Returns information about one or more pull request events}
\usage{
codecommit_describe_pull_request_events(pullRequestId,
  pullRequestEventType, actorArn, nextToken, maxResults)
}
\arguments{
\item{pullRequestId}{[required] The system-generated ID of the pull request. To get this ID, use
ListPullRequests.}

\item{pullRequestEventType}{Optional. The pull request event type about which you want to return
information.}

\item{actorArn}{The Amazon Resource Name (ARN) of the user whose actions resulted in the
event. Examples include updating the pull request with additional
commits or changing the status of a pull request.}

\item{nextToken}{An enumeration token that when provided in a request, returns the next
batch of the results.}

\item{maxResults}{A non-negative integer used to limit the number of returned results. The
default is 100 events, which is also the maximum number of events that
can be returned in a result.}
}
\description{
Returns information about one or more pull request events.
}
\section{Request syntax}{
\preformatted{codecommit$describe_pull_request_events(
  pullRequestId = "string",
  pullRequestEventType = "PULL_REQUEST_CREATED"|"PULL_REQUEST_STATUS_CHANGED"|"PULL_REQUEST_SOURCE_REFERENCE_UPDATED"|"PULL_REQUEST_MERGE_STATE_CHANGED",
  actorArn = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
