% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_commit}
\alias{codecommit_get_commit}
\title{Returns information about a commit, including commit message and
committer information}
\usage{
codecommit_get_commit(repositoryName, commitId)
}
\arguments{
\item{repositoryName}{[required] The name of the repository to which the commit was made.}

\item{commitId}{[required] The commit ID. Commit IDs are the full SHA of the commit.}
}
\description{
Returns information about a commit, including commit message and
committer information.
}
\section{Request syntax}{
\preformatted{svc$get_commit(
  repositoryName = "string",
  commitId = "string"
)
}
}

\keyword{internal}
