% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_put_repository_triggers}
\alias{codecommit_put_repository_triggers}
\title{Replaces all triggers for a repository}
\usage{
codecommit_put_repository_triggers(repositoryName, triggers)
}
\arguments{
\item{repositoryName}{[required] The name of the repository where you want to create or update the
trigger.}

\item{triggers}{[required] The JSON block of configuration information for each trigger.}
}
\description{
Replaces all triggers for a repository. This can be used to create or
delete triggers.
}
\section{Request syntax}{
\preformatted{codecommit$put_repository_triggers(
  repositoryName = "string",
  triggers = list(
    list(
      name = "string",
      destinationArn = "string",
      customData = "string",
      branches = list(
        "string"
      ),
      events = list(
        "all"|"updateReference"|"createReference"|"deleteReference"
      )
    )
  )
)
}
}

\keyword{internal}
