% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_comment}
\alias{codecommit_update_comment}
\title{Replaces the contents of a comment}
\usage{
codecommit_update_comment(commentId, content)
}
\arguments{
\item{commentId}{[required] The system-generated ID of the comment you want to update. To get this
ID, use GetCommentsForComparedCommit or GetCommentsForPullRequest.}

\item{content}{[required] The updated content with which you want to replace the existing content
of the comment.}
}
\description{
Replaces the contents of a comment.
}
\section{Request syntax}{
\preformatted{codecommit$update_comment(
  commentId = "string",
  content = "string"
)
}
}

\keyword{internal}
