% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_deregister_on_premises_instance}
\alias{codedeploy_deregister_on_premises_instance}
\title{Deregisters an on-premises instance}
\usage{
codedeploy_deregister_on_premises_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the on-premises instance to deregister.}
}
\description{
Deregisters an on-premises instance.
}
\section{Request syntax}{
\preformatted{codedeploy$deregister_on_premises_instance(
  instanceName = "string"
)
}
}

\keyword{internal}
