% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_on_premises_instance}
\alias{codedeploy_get_on_premises_instance}
\title{Gets information about an on-premises instance}
\usage{
codedeploy_get_on_premises_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the on-premises instance about which to get information.}
}
\description{
Gets information about an on-premises instance.
}
\section{Request syntax}{
\preformatted{codedeploy$get_on_premises_instance(
  instanceName = "string"
)
}
}

\keyword{internal}
