% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_add_notification_channels}
\alias{codeguruprofiler_add_notification_channels}
\title{Add up to 2 anomaly notifications channels for a profiling group}
\usage{
codeguruprofiler_add_notification_channels(channels, profilingGroupName)
}
\arguments{
\item{channels}{[required] One or 2 channels to report to when anomalies are detected.}

\item{profilingGroupName}{[required] The name of the profiling group that we are setting up notifications
for.}
}
\description{
Add up to 2 anomaly notifications channels for a profiling group.
}
\section{Request syntax}{
\preformatted{svc$add_notification_channels(
  channels = list(
    list(
      eventPublishers = list(
        "AnomalyDetection"
      ),
      id = "string",
      uri = "string"
    )
  ),
  profilingGroupName = "string"
)
}
}

\keyword{internal}
