% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_configure_agent}
\alias{codeguruprofiler_configure_agent}
\title{Provides the configuration to use for an agent of the profiling group}
\usage{
codeguruprofiler_configure_agent(fleetInstanceId, profilingGroupName)
}
\arguments{
\item{fleetInstanceId}{}

\item{profilingGroupName}{[required]}
}
\description{
Provides the configuration to use for an agent of the profiling group.
}
\section{Request syntax}{
\preformatted{svc$configure_agent(
  fleetInstanceId = "string",
  profilingGroupName = "string"
)
}
}

\keyword{internal}
