% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_create_profiling_group}
\alias{codeguruprofiler_create_profiling_group}
\title{Create a profiling group}
\usage{
codeguruprofiler_create_profiling_group(agentOrchestrationConfig,
  clientToken, profilingGroupName)
}
\arguments{
\item{agentOrchestrationConfig}{}

\item{clientToken}{[required]}

\item{profilingGroupName}{[required]}
}
\description{
Create a profiling group.
}
\section{Request syntax}{
\preformatted{svc$create_profiling_group(
  agentOrchestrationConfig = list(
    profilingEnabled = TRUE|FALSE
  ),
  clientToken = "string",
  profilingGroupName = "string"
)
}
}

\keyword{internal}
