% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_describe_profiling_group}
\alias{codeguruprofiler_describe_profiling_group}
\title{Describe a profiling group}
\usage{
codeguruprofiler_describe_profiling_group(profilingGroupName)
}
\arguments{
\item{profilingGroupName}{[required]}
}
\description{
Describe a profiling group.
}
\section{Request syntax}{
\preformatted{svc$describe_profiling_group(
  profilingGroupName = "string"
)
}
}

\keyword{internal}
