% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_get_notification_configuration}
\alias{codeguruprofiler_get_notification_configuration}
\title{Get the current configuration for anomaly notifications for a profiling
group}
\usage{
codeguruprofiler_get_notification_configuration(profilingGroupName)
}
\arguments{
\item{profilingGroupName}{[required] The name of the profiling group we want to get the notification
configuration for.}
}
\description{
Get the current configuration for anomaly notifications for a profiling
group.
}
\section{Request syntax}{
\preformatted{svc$get_notification_configuration(
  profilingGroupName = "string"
)
}
}

\keyword{internal}
