% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_get_profile}
\alias{codeguruprofiler_get_profile}
\title{Get the aggregated profile of a profiling group for the specified time
range}
\usage{
codeguruprofiler_get_profile(accept, endTime, maxDepth, period,
  profilingGroupName, startTime)
}
\arguments{
\item{accept}{The format of the profile to return. Supports application/json or
application/x-amzn-ion. Defaults to application/x-amzn-ion.}

\item{endTime}{The end time of the profile to get. Either period or endTime must be
specified. Must be greater than start and the overall time range to be
in the past and not larger than a week.}

\item{maxDepth}{}

\item{period}{The period of the profile to get. Exactly two of \code{startTime}, \code{period}
and \code{endTime} must be specified. Must be positive and the overall time
range to be in the past and not larger than a week.}

\item{profilingGroupName}{[required]}

\item{startTime}{The start time of the profile to get.}
}
\description{
Get the aggregated profile of a profiling group for the specified time
range. If the requested time range does not align with the available
aggregated profiles, it will be expanded to attain alignment. If
aggregated profiles are available only for part of the period requested,
the profile is returned from the earliest available to the latest within
the requested time range. For instance, if the requested time range is
from 00:00 to 00:20 and the available profiles are from 00:15 to 00:25,
then the returned profile will be from 00:15 to 00:20.
}
\section{Request syntax}{
\preformatted{svc$get_profile(
  accept = "string",
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  maxDepth = 123,
  period = "string",
  profilingGroupName = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}

\keyword{internal}
