% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_list_profile_times}
\alias{codeguruprofiler_list_profile_times}
\title{List the start times of the available aggregated profiles of a profiling
group for an aggregation period within the specified time range}
\usage{
codeguruprofiler_list_profile_times(endTime, maxResults, nextToken,
  orderBy, period, profilingGroupName, startTime)
}
\arguments{
\item{endTime}{[required] The end time of the time range to list profiles until.}

\item{maxResults}{}

\item{nextToken}{}

\item{orderBy}{The order (ascending or descending by start time of the profile) to list
the profiles by. Defaults to TIMESTAMP\\_DESCENDING.}

\item{period}{[required] The aggregation period to list the profiles for.}

\item{profilingGroupName}{[required]}

\item{startTime}{[required] The start time of the time range to list the profiles from.}
}
\description{
List the start times of the available aggregated profiles of a profiling
group for an aggregation period within the specified time range.
}
\section{Request syntax}{
\preformatted{svc$list_profile_times(
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  maxResults = 123,
  nextToken = "string",
  orderBy = "TimestampAscending"|"TimestampDescending",
  period = "P1D"|"PT1H"|"PT5M",
  profilingGroupName = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}

\keyword{internal}
