% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_list_profiling_groups}
\alias{codeguruprofiler_list_profiling_groups}
\title{List profiling groups in the account}
\usage{
codeguruprofiler_list_profiling_groups(includeDescription, maxResults,
  nextToken)
}
\arguments{
\item{includeDescription}{If set to true, returns the full description of the profiling groups
instead of the names. Defaults to false.}

\item{maxResults}{}

\item{nextToken}{}
}
\description{
List profiling groups in the account.
}
\section{Request syntax}{
\preformatted{svc$list_profiling_groups(
  includeDescription = TRUE|FALSE,
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
