% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_list_tags_for_resource}
\alias{codeguruprofiler_list_tags_for_resource}
\title{Returns a list of the tags that are assigned to a specified resource}
\usage{
codeguruprofiler_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that contains the tags to
return.}
}
\description{
Returns a list of the tags that are assigned to a specified resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
