% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_remove_permission}
\alias{codeguruprofiler_remove_permission}
\title{Removes permissions from a profiling group's resource-based policy that
are provided using an action group}
\usage{
codeguruprofiler_remove_permission(actionGroup, profilingGroupName,
  revisionId)
}
\arguments{
\item{actionGroup}{[required] Specifies an action group that contains the permissions to remove from a
profiling group's resource-based policy. One action group is supported,
\code{agentPermissions}, which grants \code{ConfigureAgent} and \code{PostAgentProfile}
permissions.}

\item{profilingGroupName}{[required] The name of the profiling group.}

\item{revisionId}{[required] A universally unique identifier (UUID) for the revision of the
resource-based policy from which you want to remove permissions.}
}
\description{
Removes permissions from a profiling group's resource-based policy that
are provided using an action group. The one supported action group that
can be removed is \code{agentPermission} which grants \code{ConfigureAgent} and
\code{PostAgent} permissions. For more information, see \href{https://docs.aws.amazon.com/codeguru/latest/profiler-ug/resource-based-policies.html}{Resource-based policies in CodeGuru Profiler}
in the \emph{Amazon CodeGuru Profiler User Guide},
\href{https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_ConfigureAgent.html}{\code{ConfigureAgent}}
, and
\href{https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_PostAgentProfile.html}{\code{PostAgentProfile}}
.
}
\section{Request syntax}{
\preformatted{svc$remove_permission(
  actionGroup = "agentPermissions",
  profilingGroupName = "string",
  revisionId = "string"
)
}
}

\keyword{internal}
