% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_associate_repository}
\alias{codegurureviewer_associate_repository}
\title{Associates an AWS CodeCommit repository with Amazon CodeGuru Reviewer}
\usage{
codegurureviewer_associate_repository(Repository, ClientRequestToken)
}
\arguments{
\item{Repository}{[required] The repository to associate.}

\item{ClientRequestToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.

If you want to add a new repository association, this parameter
specifies a unique identifier for the new repository association that
helps ensure idempotency.

If you use the AWS CLI or one of the AWS SDK to call this operation,
then you can leave this parameter empty. The CLI or SDK generates a
random UUID for you and includes that in the request. If you don\'t use
the SDK and instead generate a raw HTTP request to the Secrets Manager
service endpoint, then you must generate a ClientRequestToken yourself
for new versions and include that value in the request.

You typically only need to interact with this value if you implement
your own retry logic and want to ensure that a given repository
association is not created twice. We recommend that you generate a
UUID-type value to ensure uniqueness within the specified repository
association.

Amazon CodeGuru Reviewer uses this value to prevent the accidental
creation of duplicate repository associations if there are failures and
retries.}
}
\description{
Associates an AWS CodeCommit repository with Amazon CodeGuru Reviewer.
When you associate an AWS CodeCommit repository with Amazon CodeGuru
Reviewer, Amazon CodeGuru Reviewer will provide recommendations for each
pull request. You can view recommendations in the AWS CodeCommit
repository.
}
\details{
You can associate a GitHub repository using the Amazon CodeGuru Reviewer
console.
}
\section{Request syntax}{
\preformatted{svc$associate_repository(
  Repository = list(
    CodeCommit = list(
      Name = "string"
    )
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
