% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_describe_repository_association}
\alias{codegurureviewer_describe_repository_association}
\title{Describes a repository association}
\usage{
codegurureviewer_describe_repository_association(AssociationArn)
}
\arguments{
\item{AssociationArn}{[required] The Amazon Resource Name (ARN) identifying the association.}
}
\description{
Describes a repository association.
}
\section{Request syntax}{
\preformatted{svc$describe_repository_association(
  AssociationArn = "string"
)
}
}

\keyword{internal}
