% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_disassociate_repository}
\alias{codegurureviewer_disassociate_repository}
\title{Removes the association between Amazon CodeGuru Reviewer and a
repository}
\usage{
codegurureviewer_disassociate_repository(AssociationArn)
}
\arguments{
\item{AssociationArn}{[required] The Amazon Resource Name (ARN) identifying the association.}
}
\description{
Removes the association between Amazon CodeGuru Reviewer and a
repository.
}
\section{Request syntax}{
\preformatted{svc$disassociate_repository(
  AssociationArn = "string"
)
}
}

\keyword{internal}
