% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_list_repository_associations}
\alias{codegurureviewer_list_repository_associations}
\title{Lists repository associations}
\usage{
codegurureviewer_list_repository_associations(ProviderTypes, States,
  Names, Owners, MaxResults, NextToken)
}
\arguments{
\item{ProviderTypes}{List of provider types to use as a filter.}

\item{States}{List of states to use as a filter.}

\item{Names}{List of names to use as a filter.}

\item{Owners}{List of owners to use as a filter. For AWS CodeCommit, the owner is the
AWS account id. For GitHub, it is the GitHub account name.}

\item{MaxResults}{The maximum number of repository association results returned by
\code{ListRepositoryAssociations} in paginated output. When this parameter is
used, \code{ListRepositoryAssociations} only returns \code{maxResults} results in
a single page along with a \code{nextToken} response element. The remaining
results of the initial request can be seen by sending another
\code{ListRepositoryAssociations} request with the returned \code{nextToken}
value. This value can be between 1 and 100. If this parameter is not
used, then \code{ListRepositoryAssociations} returns up to 100 results and a
\code{nextToken} value if applicable.}

\item{NextToken}{The \code{nextToken} value returned from a previous paginated
\code{ListRepositoryAssociations} request where \code{maxResults} was used and the
results exceeded the value of that parameter. Pagination continues from
the end of the previous results that returned the \code{nextToken} value.

This token should be treated as an opaque identifier that is only used
to retrieve the next items in a list and not for other programmatic
purposes.}
}
\description{
Lists repository associations. You can optionally filter on one or more
of the following recommendation properties: provider types, states,
names, and owners.
}
\section{Request syntax}{
\preformatted{svc$list_repository_associations(
  ProviderTypes = list(
    "CodeCommit"|"GitHub"
  ),
  States = list(
    "Associated"|"Associating"|"Failed"|"Disassociating"
  ),
  Names = list(
    "string"
  ),
  Owners = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
