% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_acknowledge_job}
\alias{codepipeline_acknowledge_job}
\title{Returns information about a specified job and whether that job has been
received by the job worker}
\usage{
codepipeline_acknowledge_job(jobId, nonce)
}
\arguments{
\item{jobId}{[required] The unique system-generated ID of the job for which you want to confirm
receipt.}

\item{nonce}{[required] A system-generated random number that AWS CodePipeline uses to ensure
that the job is being worked on by only one job worker. Get this number
from the response of the PollForJobs request that returned this job.}
}
\description{
Returns information about a specified job and whether that job has been
received by the job worker. Only used for custom actions.
}
\section{Request syntax}{
\preformatted{codepipeline$acknowledge_job(
  jobId = "string",
  nonce = "string"
)
}
}

\keyword{internal}
