% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_list_pipelines}
\alias{codepipeline_list_pipelines}
\title{Gets a summary of all of the pipelines associated with your account}
\usage{
codepipeline_list_pipelines(nextToken)
}
\arguments{
\item{nextToken}{An identifier that was returned from the previous list pipelines call.
It can be used to return the next set of pipelines in the list.}
}
\description{
Gets a summary of all of the pipelines associated with your account.
}
\section{Request syntax}{
\preformatted{svc$list_pipelines(
  nextToken = "string"
)
}
}

\keyword{internal}
