% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_put_job_failure_result}
\alias{codepipeline_put_job_failure_result}
\title{Represents the failure of a job as returned to the pipeline by a job
worker}
\usage{
codepipeline_put_job_failure_result(jobId, failureDetails)
}
\arguments{
\item{jobId}{[required] The unique system-generated ID of the job that failed. This is the same
ID returned from PollForJobs.}

\item{failureDetails}{[required] The details about the failure of a job.}
}
\description{
Represents the failure of a job as returned to the pipeline by a job
worker. Only used for custom actions.
}
\section{Request syntax}{
\preformatted{codepipeline$put_job_failure_result(
  jobId = "string",
  failureDetails = list(
    type = "JobFailed"|"ConfigurationError"|"PermissionError"|"RevisionOutOfSync"|"RevisionUnavailable"|"SystemUnavailable",
    message = "string",
    externalExecutionId = "string"
  )
)
}
}

\keyword{internal}
