% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_service.R
\name{codestar}
\alias{codestar}
\title{AWS CodeStar}
\usage{
codestar(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
This is the API reference for AWS CodeStar. This reference provides
descriptions of the operations and data types for the AWS CodeStar API
along with usage examples.

You can use the AWS CodeStar API to work with:

Projects and their resources, by calling the following:
\itemize{
\item \code{\link[=codestar_delete_project]{delete_project}}, which deletes a
project.
\item \code{\link[=codestar_describe_project]{describe_project}}, which lists the
attributes of a project.
\item \code{\link[=codestar_list_projects]{list_projects}}, which lists all projects
associated with your AWS account.
\item \code{\link[=codestar_list_resources]{list_resources}}, which lists the
resources associated with a project.
\item \code{\link[=codestar_list_tags_for_project]{list_tags_for_project}}, which
lists the tags associated with a project.
\item \code{\link[=codestar_tag_project]{tag_project}}, which adds tags to a project.
\item \code{\link[=codestar_untag_project]{untag_project}}, which removes tags from a
project.
\item \code{\link[=codestar_update_project]{update_project}}, which updates the
attributes of a project.
}

Teams and team members, by calling the following:
\itemize{
\item \code{\link[=codestar_associate_team_member]{associate_team_member}}, which
adds an IAM user to the team for a project.
\item \code{\link[=codestar_disassociate_team_member]{disassociate_team_member}},
which removes an IAM user from the team for a project.
\item \code{\link[=codestar_list_team_members]{list_team_members}}, which lists all
the IAM users in the team for a project, including their roles and
attributes.
\item \code{\link[=codestar_update_team_member]{update_team_member}}, which updates a
team member's attributes in a project.
}

Users, by calling the following:
\itemize{
\item \code{\link[=codestar_create_user_profile]{create_user_profile}}, which creates
a user profile that contains data associated with the user across
all projects.
\item \code{\link[=codestar_delete_user_profile]{delete_user_profile}}, which deletes
all user profile information across all projects.
\item \code{\link[=codestar_describe_user_profile]{describe_user_profile}}, which
describes the profile of a user.
\item \code{\link[=codestar_list_user_profiles]{list_user_profiles}}, which lists all
user profiles.
\item \code{\link[=codestar_update_user_profile]{update_user_profile}}, which updates
the profile for a user.
}
}
\section{Service syntax}{
\preformatted{svc <- codestar(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=codestar_associate_team_member]{associate_team_member} \tab Adds an IAM user to the team for an AWS CodeStar project\cr
\link[=codestar_create_project]{create_project} \tab Creates a project, including project resources\cr
\link[=codestar_create_user_profile]{create_user_profile} \tab Creates a profile for a user that includes user preferences, such as the display name and email address assocciated with the user, in AWS CodeStar\cr
\link[=codestar_delete_project]{delete_project} \tab Deletes a project, including project resources\cr
\link[=codestar_delete_user_profile]{delete_user_profile} \tab Deletes a user profile in AWS CodeStar, including all personal preference data associated with that profile, such as display name and email address\cr
\link[=codestar_describe_project]{describe_project} \tab Describes a project and its resources\cr
\link[=codestar_describe_user_profile]{describe_user_profile} \tab Describes a user in AWS CodeStar and the user attributes across all projects\cr
\link[=codestar_disassociate_team_member]{disassociate_team_member} \tab Removes a user from a project\cr
\link[=codestar_list_projects]{list_projects} \tab Lists all projects in AWS CodeStar associated with your AWS account\cr
\link[=codestar_list_resources]{list_resources} \tab Lists resources associated with a project in AWS CodeStar\cr
\link[=codestar_list_tags_for_project]{list_tags_for_project} \tab Gets the tags for a project\cr
\link[=codestar_list_team_members]{list_team_members} \tab Lists all team members associated with a project\cr
\link[=codestar_list_user_profiles]{list_user_profiles} \tab Lists all the user profiles configured for your AWS account in AWS CodeStar\cr
\link[=codestar_tag_project]{tag_project} \tab Adds tags to a project\cr
\link[=codestar_untag_project]{untag_project} \tab Removes tags from a project\cr
\link[=codestar_update_project]{update_project} \tab Updates a project in AWS CodeStar\cr
\link[=codestar_update_team_member]{update_team_member} \tab Updates a team member's attributes in an AWS CodeStar project\cr
\link[=codestar_update_user_profile]{update_user_profile} \tab Updates a user's profile in AWS CodeStar
}
}

\examples{
\dontrun{
svc <- codestar()
svc$associate_team_member(
  Foo = 123
)
}

}
