% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_create_connection}
\alias{codestarconnections_create_connection}
\title{Creates a connection that can then be given to other AWS services like
CodePipeline so that it can access third-party code repositories}
\usage{
codestarconnections_create_connection(ProviderType, ConnectionName)
}
\arguments{
\item{ProviderType}{[required] The name of the external provider where your third-party code repository
is configured. Currently, the valid provider type is Bitbucket.}

\item{ConnectionName}{[required] The name of the connection to be created. The name must be unique in the
calling AWS account.}
}
\description{
Creates a connection that can then be given to other AWS services like
CodePipeline so that it can access third-party code repositories. The
connection is in pending status until the third-party connection
handshake is completed from the console.
}
\section{Request syntax}{
\preformatted{svc$create_connection(
  ProviderType = "Bitbucket",
  ConnectionName = "string"
)
}
}

\keyword{internal}
