% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_get_connection}
\alias{codestarconnections_get_connection}
\title{Returns the connection ARN and details such as status, owner, and
provider type}
\usage{
codestarconnections_get_connection(ConnectionArn)
}
\arguments{
\item{ConnectionArn}{[required] The Amazon Resource Name (ARN) of a connection.}
}
\description{
Returns the connection ARN and details such as status, owner, and
provider type.
}
\section{Request syntax}{
\preformatted{svc$get_connection(
  ConnectionArn = "string"
)
}
}

\keyword{internal}
