% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_get_host}
\alias{codestarconnections_get_host}
\title{Returns the host ARN and details such as status, provider type,
endpoint, and, if applicable, the VPC configuration}
\usage{
codestarconnections_get_host(HostArn)
}
\arguments{
\item{HostArn}{[required] The Amazon Resource Name (ARN) of the requested host.}
}
\description{
Returns the host ARN and details such as status, provider type,
endpoint, and, if applicable, the VPC configuration.
}
\section{Request syntax}{
\preformatted{svc$get_host(
  HostArn = "string"
)
}
}

\keyword{internal}
