% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_list_hosts}
\alias{codestarconnections_list_hosts}
\title{Lists the hosts associated with your account}
\usage{
codestarconnections_list_hosts(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token that was returned from the previous \code{ListHosts} call, which
can be used to return the next set of hosts in the list.}
}
\description{
Lists the hosts associated with your account.
}
\section{Request syntax}{
\preformatted{svc$list_hosts(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
