% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_update_host}
\alias{codestarconnections_update_host}
\title{Updates a specified host with the provided configurations}
\usage{
codestarconnections_update_host(HostArn, ProviderEndpoint,
  VpcConfiguration)
}
\arguments{
\item{HostArn}{[required] The Amazon Resource Name (ARN) of the host to be updated.}

\item{ProviderEndpoint}{The URL or endpoint of the host to be updated.}

\item{VpcConfiguration}{The VPC configuration of the host to be updated. A VPC must be
configured and the infrastructure to be represented by the host must
already be connected to the VPC.}
}
\description{
Updates a specified host with the provided configurations.
}
\section{Request syntax}{
\preformatted{svc$update_host(
  HostArn = "string",
  ProviderEndpoint = "string",
  VpcConfiguration = list(
    VpcId = "string",
    SubnetIds = list(
      "string"
    ),
    SecurityGroupIds = list(
      "string"
    ),
    TlsCertificate = "string"
  )
)
}
}

\keyword{internal}
