% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_delete_target}
\alias{codestarnotifications_delete_target}
\title{Deletes a specified target for notifications}
\usage{
codestarnotifications_delete_target(TargetAddress, ForceUnsubscribeAll)
}
\arguments{
\item{TargetAddress}{[required] The Amazon Resource Name (ARN) of the SNS topic to delete.}

\item{ForceUnsubscribeAll}{A Boolean value that can be used to delete all associations with this
SNS topic. The default value is FALSE. If set to TRUE, all associations
between that target and every notification rule in your AWS account are
deleted.}
}
\description{
Deletes a specified target for notifications.
}
\section{Request syntax}{
\preformatted{svc$delete_target(
  TargetAddress = "string",
  ForceUnsubscribeAll = TRUE|FALSE
)
}
}

\keyword{internal}
