% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_delete_identity_pool}
\alias{cognitoidentity_delete_identity_pool}
\title{Deletes an identity pool}
\usage{
cognitoidentity_delete_identity_pool(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{[required] An identity pool ID in the format REGION:GUID.}
}
\description{
Deletes an identity pool. Once a pool is deleted, users will not be able
to authenticate with the pool.
}
\details{
You must use AWS Developer credentials to call this API.
}
\section{Request syntax}{
\preformatted{cognitoidentity$delete_identity_pool(
  IdentityPoolId = "string"
)
}
}

\keyword{internal}
