% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_update_identity_pool}
\alias{cognitoidentity_update_identity_pool}
\title{Updates a user pool}
\usage{
cognitoidentity_update_identity_pool(IdentityPoolId, IdentityPoolName,
  AllowUnauthenticatedIdentities, SupportedLoginProviders,
  DeveloperProviderName, OpenIdConnectProviderARNs,
  CognitoIdentityProviders, SamlProviderARNs)
}
\arguments{
\item{IdentityPoolId}{[required] An identity pool ID in the format REGION:GUID.}

\item{IdentityPoolName}{[required] A string that you provide.}

\item{AllowUnauthenticatedIdentities}{[required] TRUE if the identity pool supports unauthenticated logins.}

\item{SupportedLoginProviders}{Optional key:value pairs mapping provider names to provider app IDs.}

\item{DeveloperProviderName}{The "domain" by which Cognito will refer to your users.}

\item{OpenIdConnectProviderARNs}{A list of OpendID Connect provider ARNs.}

\item{CognitoIdentityProviders}{A list representing an Amazon Cognito Identity User Pool and its client
ID.}

\item{SamlProviderARNs}{An array of Amazon Resource Names (ARNs) of the SAML provider for your
identity pool.}
}
\description{
Updates a user pool.
}
\details{
You must use AWS Developer credentials to call this API.
}
\section{Request syntax}{
\preformatted{cognitoidentity$update_identity_pool(
  IdentityPoolId = "string",
  IdentityPoolName = "string",
  AllowUnauthenticatedIdentities = TRUE|FALSE,
  SupportedLoginProviders = list(
    "string"
  ),
  DeveloperProviderName = "string",
  OpenIdConnectProviderARNs = list(
    "string"
  ),
  CognitoIdentityProviders = list(
    list(
      ProviderName = "string",
      ClientId = "string",
      ServerSideTokenCheck = TRUE|FALSE
    )
  ),
  SamlProviderARNs = list(
    "string"
  )
)
}
}

\keyword{internal}
