% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_get_user}
\alias{cognitoidentityprovider_admin_get_user}
\title{Gets the specified user by user name in a user pool as an administrator}
\usage{
cognitoidentityprovider_admin_get_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to get information
about the user.}

\item{Username}{[required] The user name of the user you wish to retrieve.}
}
\description{
Gets the specified user by user name in a user pool as an administrator.
Works on any user.
}
\details{
Calling this action requires developer credentials.
}
\section{Request syntax}{
\preformatted{svc$admin_get_user(
  UserPoolId = "string",
  Username = "string"
)
}
}

\keyword{internal}
