% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_set_user_password}
\alias{cognitoidentityprovider_admin_set_user_password}
\title{Admin set user password}
\usage{
cognitoidentityprovider_admin_set_user_password(UserPoolId, Username,
  Password, Permanent)
}
\arguments{
\item{UserPoolId}{[required]}

\item{Username}{[required]}

\item{Password}{[required]}

\item{Permanent}{}
}
\description{
Admin set user password
}
\section{Request syntax}{
\preformatted{svc$admin_set_user_password(
  UserPoolId = "string",
  Username = "string",
  Password = "string",
  Permanent = TRUE|FALSE
)
}
}

\keyword{internal}
