% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_create_group}
\alias{cognitoidentityprovider_create_group}
\title{Creates a new group in the specified user pool}
\usage{
cognitoidentityprovider_create_group(GroupName, UserPoolId, Description,
  RoleArn, Precedence)
}
\arguments{
\item{GroupName}{[required] The name of the group. Must be unique.}

\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{Description}{A string containing the description of the group.}

\item{RoleArn}{The role ARN for the group.}

\item{Precedence}{A nonnegative integer value that specifies the precedence of this group
relative to the other groups that a user can belong to in the user pool.
Zero is the highest precedence value. Groups with lower \code{Precedence}
values take precedence over groups with higher or null \code{Precedence}
values. If a user belongs to two or more groups, it is the group with
the lowest precedence value whose role ARN will be used in the
\code{cognito:roles} and \code{cognito:preferred_role} claims in the user\'s
tokens.

Two groups can have the same \code{Precedence} value. If this happens,
neither group takes precedence over the other. If two groups with the
same \code{Precedence} have the same role ARN, that role is used in the
\code{cognito:preferred_role} claim in tokens for users in each group. If the
two groups have different role ARNs, the \code{cognito:preferred_role} claim
is not set in users\' tokens.

The default \code{Precedence} value is null.}
}
\description{
Creates a new group in the specified user pool.
}
\details{
Calling this action requires developer credentials.
}
\section{Request syntax}{
\preformatted{svc$create_group(
  GroupName = "string",
  UserPoolId = "string",
  Description = "string",
  RoleArn = "string",
  Precedence = 123
)
}
}

\keyword{internal}
