% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_group}
\alias{cognitoidentityprovider_delete_group}
\title{Deletes a group}
\usage{
cognitoidentityprovider_delete_group(GroupName, UserPoolId)
}
\arguments{
\item{GroupName}{[required] The name of the group.}

\item{UserPoolId}{[required] The user pool ID for the user pool.}
}
\description{
Deletes a group. Currently only groups with no members can be deleted.
}
\details{
Requires developer credentials.
}
\section{Request syntax}{
\preformatted{svc$delete_group(
  GroupName = "string",
  UserPoolId = "string"
)
}
}

\keyword{internal}
