% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_identity_provider}
\alias{cognitoidentityprovider_delete_identity_provider}
\title{Deletes an identity provider for a user pool}
\usage{
cognitoidentityprovider_delete_identity_provider(UserPoolId,
  ProviderName)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{ProviderName}{[required] The identity provider name.}
}
\description{
Deletes an identity provider for a user pool.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$delete_identity_provider(
  UserPoolId = "string",
  ProviderName = "string"
)
}
}

\keyword{internal}
