% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_group}
\alias{cognitoidentityprovider_get_group}
\title{Gets a group}
\usage{
cognitoidentityprovider_get_group(GroupName, UserPoolId)
}
\arguments{
\item{GroupName}{[required] The name of the group.}

\item{UserPoolId}{[required] The user pool ID for the user pool.}
}
\description{
Gets a group.
}
\details{
Requires developer credentials.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$get_group(
  GroupName = "string",
  UserPoolId = "string"
)
}
}

\keyword{internal}
