% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_user}
\alias{cognitoidentityprovider_get_user}
\title{Gets the user attributes and metadata for a user}
\usage{
cognitoidentityprovider_get_user(AccessToken)
}
\arguments{
\item{AccessToken}{[required] The access token returned by the server response to get information
about the user.}
}
\description{
Gets the user attributes and metadata for a user.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$get_user(
  AccessToken = "string"
)
}
}

\keyword{internal}
