% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_global_sign_out}
\alias{cognitoidentityprovider_global_sign_out}
\title{Signs out users from all devices}
\usage{
cognitoidentityprovider_global_sign_out(AccessToken)
}
\arguments{
\item{AccessToken}{[required] The access token.}
}
\description{
Signs out users from all devices.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$global_sign_out(
  AccessToken = "string"
)
}
}

\keyword{internal}
