% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_update_device_status}
\alias{cognitoidentityprovider_update_device_status}
\title{Updates the device status}
\usage{
cognitoidentityprovider_update_device_status(AccessToken, DeviceKey,
  DeviceRememberedStatus)
}
\arguments{
\item{AccessToken}{[required] The access token.}

\item{DeviceKey}{[required] The device key.}

\item{DeviceRememberedStatus}{The status of whether a device is remembered.}
}
\description{
Updates the device status.
}
\section{Request syntax}{
\preformatted{svc$update_device_status(
  AccessToken = "string",
  DeviceKey = "string",
  DeviceRememberedStatus = "remembered"|"not_remembered"
)
}
}

\keyword{internal}
