% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_get_cognito_events}
\alias{cognitosync_get_cognito_events}
\title{Gets the events and the corresponding Lambda functions associated with
an identity pool}
\usage{
cognitosync_get_cognito_events(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{[required] The Cognito Identity Pool ID for the request}
}
\description{
Gets the events and the corresponding Lambda functions associated with
an identity pool.
}
\details{
This API can only be called with developer credentials. You cannot call
this API with the temporary user credentials provided by Cognito
Identity.
}
\section{Request syntax}{
\preformatted{cognitosync$get_cognito_events(
  IdentityPoolId = "string"
)
}
}

\keyword{internal}
