% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_list_document_classifiers}
\alias{comprehend_list_document_classifiers}
\title{Gets a list of the document classifiers that you have created}
\usage{
comprehend_list_document_classifiers(Filter, NextToken, MaxResults)
}
\arguments{
\item{Filter}{Filters the jobs that are returned. You can filter jobs on their name,
status, or the date and time that they were submitted. You can only set
one filter at a time.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{The maximum number of results to return in each page. The default is
100.}
}
\description{
Gets a list of the document classifiers that you have created.
}
\section{Request syntax}{
\preformatted{comprehend$list_document_classifiers(
  Filter = list(
    Status = "SUBMITTED"|"TRAINING"|"DELETING"|"STOP_REQUESTED"|"STOPPED"|"IN_ERROR"|"TRAINED",
    SubmitTimeBefore = as.POSIXct(
      "2015-01-01"
    ),
    SubmitTimeAfter = as.POSIXct(
      "2015-01-01"
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
