% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_describe_rx_norm_inference_job}
\alias{comprehendmedical_describe_rx_norm_inference_job}
\title{Gets the properties associated with an InferRxNorm job}
\usage{
comprehendmedical_describe_rx_norm_inference_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Amazon Comprehend Medical generated for the job. The
StartRxNormInferenceJob operation returns this identifier in its
response.}
}
\description{
Gets the properties associated with an InferRxNorm job. Use this
operation to get the status of an inference job.
}
\section{Request syntax}{
\preformatted{svc$describe_rx_norm_inference_job(
  JobId = "string"
)
}
}

\keyword{internal}
