% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_list_rx_norm_inference_jobs}
\alias{comprehendmedical_list_rx_norm_inference_jobs}
\title{Gets a list of InferRxNorm jobs that you have submitted}
\usage{
comprehendmedical_list_rx_norm_inference_jobs(Filter, NextToken,
  MaxResults)
}
\arguments{
\item{Filter}{Filters the jobs that are returned. You can filter jobs based on their
names, status, or the date and time that they were submitted. You can
only set one filter at a time.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{Identifies the next page of results to return.}
}
\description{
Gets a list of InferRxNorm jobs that you have submitted.
}
\section{Request syntax}{
\preformatted{svc$list_rx_norm_inference_jobs(
  Filter = list(
    JobName = "string",
    JobStatus = "SUBMITTED"|"IN_PROGRESS"|"COMPLETED"|"PARTIAL_SUCCESS"|"FAILED"|"STOP_REQUESTED"|"STOPPED",
    SubmitTimeBefore = as.POSIXct(
      "2015-01-01"
    ),
    SubmitTimeAfter = as.POSIXct(
      "2015-01-01"
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
