% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_export_ec2_instance_recommendations}
\alias{computeoptimizer_export_ec2_instance_recommendations}
\title{Exports optimization recommendations for Amazon EC2 instances}
\usage{
computeoptimizer_export_ec2_instance_recommendations(accountIds,
  filters, fieldsToExport, s3DestinationConfig, fileFormat,
  includeMemberAccounts)
}
\arguments{
\item{accountIds}{The IDs of the AWS accounts for which to export instance
recommendations.

If your account is the management account of an organization, use this
parameter to specify the member accounts for which you want to export
recommendations.

This parameter cannot be specified together with the include member
accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if
this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.}

\item{filters}{An array of objects that describe a filter to export a more specific set
of instance recommendations.}

\item{fieldsToExport}{The recommendations data to include in the export file. For more
information about the fields that can be exported, see \href{https://docs.aws.amazon.com/compute-optimizer/latest/ug/exporting-recommendations.html#exported-files}{Exported files}
in the \emph{Compute Optimizer User Guide}.}

\item{s3DestinationConfig}{[required] An object to specify the destination Amazon Simple Storage Service
(Amazon S3) bucket name and key prefix for the export job.

You must create the destination Amazon S3 bucket for your
recommendations export before you create the export job. Compute
Optimizer does not create the S3 bucket for you. After you create the S3
bucket, ensure that it has the required permission policy to allow
Compute Optimizer to write the export file to it. If you plan to specify
an object prefix when you create the export job, you must include the
object prefix in the policy that you add to the S3 bucket. For more
information, see \href{https://docs.aws.amazon.com/compute-optimizer/latest/ug/create-s3-bucket-policy-for-compute-optimizer.html}{Amazon S3 Bucket Policy for Compute Optimizer}
in the \emph{Compute Optimizer user guide}.}

\item{fileFormat}{The format of the export file.

The only export file format currently supported is \code{Csv}.}

\item{includeMemberAccounts}{Indicates whether to include recommendations for resources in all member
accounts of the organization if your account is the management account
of an organization.

The member accounts must also be opted in to Compute Optimizer.

Recommendations for member accounts of the organization are not included
in the export file if this parameter is omitted.

Recommendations for member accounts are not included in the export if
this parameter, or the account IDs parameter, is omitted.}
}
\description{
Exports optimization recommendations for Amazon EC2 instances.

Recommendations are exported in a comma-separated values (.csv) file,
and its metadata in a JavaScript Object Notation (.json) file, to an
existing Amazon Simple Storage Service (Amazon S3) bucket that you
specify. For more information, see \href{https://docs.aws.amazon.com/compute-optimizer/latest/ug/exporting-recommendations.html}{Exporting Recommendations}
in the \emph{Compute Optimizer User Guide}.

You can have only one Amazon EC2 instance export job in progress per AWS
Region.
}
\section{Request syntax}{
\preformatted{svc$export_ec2_instance_recommendations(
  accountIds = list(
    "string"
  ),
  filters = list(
    list(
      name = "Finding"|"RecommendationSourceType",
      values = list(
        "string"
      )
    )
  ),
  fieldsToExport = list(
    "AccountId"|"InstanceArn"|"InstanceName"|"Finding"|"LookbackPeriodInDays"|"CurrentInstanceType"|"UtilizationMetricsCpuMaximum"|"UtilizationMetricsMemoryMaximum"|"UtilizationMetricsEbsReadOpsPerSecondMaximum"|"UtilizationMetricsEbsWriteOpsPerSecondMaximum"|"UtilizationMetricsEbsReadBytesPerSecondMaximum"|"UtilizationMetricsEbsWriteBytesPerSecondMaximum"|"CurrentOnDemandPrice"|"CurrentStandardOneYearNoUpfrontReservedPrice"|"CurrentStandardThreeYearNoUpfrontReservedPrice"|"CurrentVCpus"|"CurrentMemory"|"CurrentStorage"|"CurrentNetwork"|"RecommendationOptionsInstanceType"|"RecommendationOptionsProjectedUtilizationMetricsCpuMaximum"|"RecommendationOptionsProjectedUtilizationMetricsMemoryMaximum"|"RecommendationOptionsPerformanceRisk"|"RecommendationOptionsVcpus"|"RecommendationOptionsMemory"|"RecommendationOptionsStorage"|"RecommendationOptionsNetwork"|"RecommendationOptionsOnDemandPrice"|"RecommendationOptionsStandardOneYearNoUpfrontReservedPrice"|"RecommendationOptionsStandardThreeYearNoUpfrontReservedPrice"|"RecommendationsSourcesRecommendationSourceArn"|"RecommendationsSourcesRecommendationSourceType"|"LastRefreshTimestamp"
  ),
  s3DestinationConfig = list(
    bucket = "string",
    keyPrefix = "string"
  ),
  fileFormat = "Csv",
  includeMemberAccounts = TRUE|FALSE
)
}
}

\keyword{internal}
