% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_auto_scaling_group_recommendations}
\alias{computeoptimizer_get_auto_scaling_group_recommendations}
\title{Returns Auto Scaling group recommendations}
\usage{
computeoptimizer_get_auto_scaling_group_recommendations(accountIds,
  autoScalingGroupArns, nextToken, maxResults, filters)
}
\arguments{
\item{accountIds}{The IDs of the AWS accounts for which to return Auto Scaling group
recommendations.

If your account is the master account of an organization, use this
parameter to specify the member accounts for which you want to return
Auto Scaling group recommendations.

Only one account ID can be specified per request.}

\item{autoScalingGroupArns}{The Amazon Resource Name (ARN) of the Auto Scaling groups for which to
return recommendations.}

\item{nextToken}{The token to advance to the next page of Auto Scaling group
recommendations.}

\item{maxResults}{The maximum number of Auto Scaling group recommendations to return with
a single request.

To retrieve the remaining results, make another request with the
returned \code{NextToken} value.}

\item{filters}{An array of objects that describe a filter that returns a more specific
list of Auto Scaling group recommendations.}
}
\description{
Returns Auto Scaling group recommendations.
}
\details{
AWS Compute Optimizer currently generates recommendations for Auto
Scaling groups that are configured to run instances of the M, C, R, T,
and X instance families. The service does not generate recommendations
for Auto Scaling groups that have a scaling policy attached to them, or
that do not have the same values for desired, minimum, and maximum
capacity. In order for Compute Optimizer to analyze your Auto Scaling
groups, they must be of a fixed size. For more information, see the \href{https://docs.aws.amazon.com/compute-optimizer/latest/ug/what-is.html}{AWS Compute Optimizer User Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_auto_scaling_group_recommendations(
  accountIds = list(
    "string"
  ),
  autoScalingGroupArns = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "Finding"|"RecommendationSourceType",
      values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
