% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_ebs_volume_recommendations}
\alias{computeoptimizer_get_ebs_volume_recommendations}
\title{Returns Amazon Elastic Block Store (Amazon EBS) volume recommendations}
\usage{
computeoptimizer_get_ebs_volume_recommendations(volumeArns, nextToken,
  maxResults, filters, accountIds)
}
\arguments{
\item{volumeArns}{The Amazon Resource Name (ARN) of the volumes for which to return
recommendations.}

\item{nextToken}{The token to advance to the next page of volume recommendations.}

\item{maxResults}{The maximum number of volume recommendations to return with a single
request.

To retrieve the remaining results, make another request with the
returned \code{NextToken} value.}

\item{filters}{An array of objects that describe a filter that returns a more specific
list of volume recommendations.}

\item{accountIds}{The IDs of the AWS accounts for which to return volume recommendations.

If your account is the management account of an organization, use this
parameter to specify the member accounts for which you want to return
volume recommendations.

Only one account ID can be specified per request.}
}
\description{
Returns Amazon Elastic Block Store (Amazon EBS) volume recommendations.
}
\details{
AWS Compute Optimizer generates recommendations for Amazon EBS volumes
that meet a specific set of requirements. For more information, see the
\href{https://docs.aws.amazon.com/compute-optimizer/latest/ug/requirements.html}{Supported resources and requirements}
in the \emph{AWS Compute Optimizer User Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_ebs_volume_recommendations(
  volumeArns = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "Finding",
      values = list(
        "string"
      )
    )
  ),
  accountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
