% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_ec2_instance_recommendations}
\alias{computeoptimizer_get_ec2_instance_recommendations}
\title{Returns Amazon EC2 instance recommendations}
\usage{
computeoptimizer_get_ec2_instance_recommendations(instanceArns,
  nextToken, maxResults, filters, accountIds)
}
\arguments{
\item{instanceArns}{The Amazon Resource Name (ARN) of the instances for which to return
recommendations.}

\item{nextToken}{The token to advance to the next page of instance recommendations.}

\item{maxResults}{The maximum number of instance recommendations to return with a single
call.

To retrieve the remaining results, make another call with the returned
\code{NextToken} value.}

\item{filters}{An array of objects that describe a filter that returns a more specific
list of instance recommendations.}

\item{accountIds}{The AWS account IDs for which to return instance recommendations.

Only one account ID can be specified per request.}
}
\description{
Returns Amazon EC2 instance recommendations.
}
\details{
AWS Compute Optimizer currently generates recommendations for Amazon
Elastic Compute Cloud (Amazon EC2) and Amazon EC2 Auto Scaling. It
generates recommendations for M, C, R, T, and X instance families. For
more information, see the \href{https://docs.aws.amazon.com/compute-optimizer/latest/ug/what-is.html}{AWS Compute Optimizer User Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_ec2_instance_recommendations(
  instanceArns = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "Finding"|"RecommendationSourceType",
      values = list(
        "string"
      )
    )
  ),
  accountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
