% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_enrollment_status}
\alias{computeoptimizer_get_enrollment_status}
\title{Returns the enrollment (opt in) status of an account to the AWS Compute
Optimizer service}
\usage{
computeoptimizer_get_enrollment_status()
}
\description{
Returns the enrollment (opt in) status of an account to the AWS Compute
Optimizer service.
}
\details{
If the account is a master account of an organization, this operation
also confirms the enrollment status of member accounts within the
organization.
}
\section{Request syntax}{
\preformatted{svc$get_enrollment_status()
}
}

\keyword{internal}
