% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_lambda_function_recommendations}
\alias{computeoptimizer_get_lambda_function_recommendations}
\title{Returns AWS Lambda function recommendations}
\usage{
computeoptimizer_get_lambda_function_recommendations(functionArns,
  accountIds, filters, nextToken, maxResults)
}
\arguments{
\item{functionArns}{The Amazon Resource Name (ARN) of the functions for which to return
recommendations.

You can specify a qualified or unqualified ARN. If you specify an
unqualified ARN without a function version suffix, Compute Optimizer
will return recommendations for the latest (\verb{$LATEST}) version of the
function. If you specify a qualified ARN with a version suffix, Compute
Optimizer will return recommendations for the specified function
version. For more information about using function versions, see \href{https://docs.aws.amazon.com/lambda/latest/dg/configuration-versions.html#versioning-versions-using}{Using versions}
in the \emph{AWS Lambda Developer Guide}.}

\item{accountIds}{The IDs of the AWS accounts for which to return function
recommendations.

If your account is the management account of an organization, use this
parameter to specify the member accounts for which you want to return
function recommendations.

Only one account ID can be specified per request.}

\item{filters}{An array of objects that describe a filter that returns a more specific
list of function recommendations.}

\item{nextToken}{The token to advance to the next page of function recommendations.}

\item{maxResults}{The maximum number of function recommendations to return with a single
request.

To retrieve the remaining results, make another request with the
returned \code{NextToken} value.}
}
\description{
Returns AWS Lambda function recommendations.

AWS Compute Optimizer generates recommendations for functions that meet
a specific set of requirements. For more information, see the \href{https://docs.aws.amazon.com/compute-optimizer/latest/ug/requirements.html}{Supported resources and requirements}
in the \emph{AWS Compute Optimizer User Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_lambda_function_recommendations(
  functionArns = list(
    "string"
  ),
  accountIds = list(
    "string"
  ),
  filters = list(
    list(
      name = "Finding"|"FindingReasonCode",
      values = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
