% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_recommendation_summaries}
\alias{computeoptimizer_get_recommendation_summaries}
\title{Returns the optimization findings for an account}
\usage{
computeoptimizer_get_recommendation_summaries(accountIds, nextToken,
  maxResults)
}
\arguments{
\item{accountIds}{The AWS account IDs for which to return recommendation summaries.

Only one account ID can be specified per request.}

\item{nextToken}{The token to advance to the next page of recommendation summaries.}

\item{maxResults}{The maximum number of recommendation summaries to return with a single
call.

To retrieve the remaining results, make another call with the returned
\code{NextToken} value.}
}
\description{
Returns the optimization findings for an account.
}
\details{
For example, it returns the number of Amazon EC2 instances in an account
that are under-provisioned, over-provisioned, or optimized. It also
returns the number of Auto Scaling groups in an account that are not
optimized, or optimized.
}
\section{Request syntax}{
\preformatted{svc$get_recommendation_summaries(
  accountIds = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
