% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_update_enrollment_status}
\alias{computeoptimizer_update_enrollment_status}
\title{Updates the enrollment (opt in) status of an account to the AWS Compute
Optimizer service}
\usage{
computeoptimizer_update_enrollment_status(status, includeMemberAccounts)
}
\arguments{
\item{status}{[required] The new enrollment status of the account.

Accepted options are \code{Active} or \code{Inactive}. You will get an error if
\code{Pending} or \code{Failed} are specified.}

\item{includeMemberAccounts}{Indicates whether to enroll member accounts of the organization if the
your account is the management account of an organization.}
}
\description{
Updates the enrollment (opt in) status of an account to the AWS Compute
Optimizer service.

If the account is a management account of an organization, this action
can also be used to enroll member accounts within the organization.
}
\section{Request syntax}{
\preformatted{svc$update_enrollment_status(
  status = "Active"|"Inactive"|"Pending"|"Failed",
  includeMemberAccounts = TRUE|FALSE
)
}
}

\keyword{internal}
