% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_evaluation_results}
\alias{configservice_delete_evaluation_results}
\title{Deletes the evaluation results for the specified AWS Config rule}
\usage{
configservice_delete_evaluation_results(ConfigRuleName)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the AWS Config rule for which you want to delete the
evaluation results.}
}
\description{
Deletes the evaluation results for the specified AWS Config rule. You
can specify one AWS Config rule per request. After you delete the
evaluation results, you can call the StartConfigRulesEvaluation API to
start evaluating your AWS resources against the rule.
}
\section{Request syntax}{
\preformatted{configservice$delete_evaluation_results(
  ConfigRuleName = "string"
)
}
}

\keyword{internal}
