% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_organization_config_rule}
\alias{configservice_delete_organization_config_rule}
\title{Delete organization config rule}
\usage{
configservice_delete_organization_config_rule(
  OrganizationConfigRuleName)
}
\arguments{
\item{OrganizationConfigRuleName}{[required]}
}
\description{
Delete organization config rule
}
\section{Request syntax}{
\preformatted{svc$delete_organization_config_rule(
  OrganizationConfigRuleName = "string"
)
}
}

\keyword{internal}
