% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_pending_aggregation_requests}
\alias{configservice_describe_pending_aggregation_requests}
\title{Returns a list of all pending aggregation requests}
\usage{
configservice_describe_pending_aggregation_requests(Limit, NextToken)
}
\arguments{
\item{Limit}{The maximum number of evaluation results returned on each page. The
default is maximum. If you specify 0, AWS Config uses the default.}

\item{NextToken}{The nextToken string returned on a previous page that you use to get the
next page of results in a paginated response.}
}
\description{
Returns a list of all pending aggregation requests.
}
\section{Request syntax}{
\preformatted{configservice$describe_pending_aggregation_requests(
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
