% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_select_resource_config}
\alias{configservice_select_resource_config}
\title{Accepts a structured query language (SQL) SELECT command, performs the
corresponding search, and returns resource configurations matching the
properties}
\usage{
configservice_select_resource_config(Expression, Limit, NextToken)
}
\arguments{
\item{Expression}{[required] The SQL query \code{SELECT} command.}

\item{Limit}{The maximum number of query results returned on each page.}

\item{NextToken}{The \code{nextToken} string returned in a previous request that you use to
request the next page of results in a paginated response.}
}
\description{
Accepts a structured query language (SQL) \code{SELECT} command, performs the
corresponding search, and returns resource configurations matching the
properties.
}
\details{
For more information about query components, see the \href{https://docs.aws.amazon.com/config/latest/developerguide/query-components.html}{\strong{Query Components}}
section in the AWS Config Developer Guide.
}
\section{Request syntax}{
\preformatted{svc$select_resource_config(
  Expression = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
