% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_lex_bot}
\alias{connect_associate_lex_bot}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_associate_lex_bot(InstanceId, LexBot)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{LexBot}{[required] The Amazon Lex box to associate with the instance.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Allows the specified Amazon Connect instance to access the specified
Amazon Lex bot.
}
\section{Request syntax}{
\preformatted{svc$associate_lex_bot(
  InstanceId = "string",
  LexBot = list(
    Name = "string",
    LexRegion = "string"
  )
)
}
}

\keyword{internal}
