% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_instance}
\alias{connect_create_instance}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_create_instance(ClientToken, IdentityManagementType,
  InstanceAlias, DirectoryId, InboundCallsEnabled, OutboundCallsEnabled)
}
\arguments{
\item{ClientToken}{The idempotency token.}

\item{IdentityManagementType}{[required] The type of identity management for your Amazon Connect users.}

\item{InstanceAlias}{The name for your instance.}

\item{DirectoryId}{The identifier for the directory.}

\item{InboundCallsEnabled}{[required] Whether your contact center handles incoming contacts.}

\item{OutboundCallsEnabled}{[required] Whether your contact center allows outbound calls.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.
}
\details{
Initiates an Amazon Connect instance with all the supported channels
enabled. It does not attach any storage (such as Amazon S3, or Kinesis)
or allow for any configurations on features such as Contact Lens for
Amazon Connect.
}
\section{Request syntax}{
\preformatted{svc$create_instance(
  ClientToken = "string",
  IdentityManagementType = "SAML"|"CONNECT_MANAGED"|"EXISTING_DIRECTORY",
  InstanceAlias = "string",
  DirectoryId = "string",
  InboundCallsEnabled = TRUE|FALSE,
  OutboundCallsEnabled = TRUE|FALSE
)
}
}

\keyword{internal}
