% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_use_case}
\alias{connect_create_use_case}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_create_use_case(InstanceId, IntegrationAssociationId,
  UseCaseType)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{IntegrationAssociationId}{[required] The identifier for the AppIntegration association.}

\item{UseCaseType}{[required] The type of use case to associate to the AppIntegration association.
Each AppIntegration association can have only one of each use case type.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Creates a use case for an AppIntegration association.
}
\section{Request syntax}{
\preformatted{svc$create_use_case(
  InstanceId = "string",
  IntegrationAssociationId = "string",
  UseCaseType = "RULES_EVALUATION"
)
}
}

\keyword{internal}
