% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_instance}
\alias{connect_describe_instance}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_instance(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Returns the current state of the specified instance identifier. It
tracks the instance while it is being created and returns an error
status if applicable.

If an instance is not created successfully, the instance status reason
field returns details relevant to the reason. The instance in a failed
state is returned only for 24 hours after the CreateInstance API was
invoked.
}
\section{Request syntax}{
\preformatted{svc$describe_instance(
  InstanceId = "string"
)
}
}

\keyword{internal}
