% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_instance_attribute}
\alias{connect_describe_instance_attribute}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_instance_attribute(InstanceId, AttributeType)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{AttributeType}{[required] The type of attribute.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.
}
\details{
Describes the specified instance attribute.
}
\section{Request syntax}{
\preformatted{svc$describe_instance_attribute(
  InstanceId = "string",
  AttributeType = "INBOUND_CALLS"|"OUTBOUND_CALLS"|"CONTACTFLOW_LOGS"|"CONTACT_LENS"|"AUTO_RESOLVE_BEST_VOICES"|"USE_CUSTOM_TTS_VOICES"|"EARLY_MEDIA"
)
}
}

\keyword{internal}
