% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_routing_profile}
\alias{connect_describe_routing_profile}
\title{Describes the specified routing profile}
\usage{
connect_describe_routing_profile(InstanceId, RoutingProfileId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}
}
\description{
Describes the specified routing profile.
}
\section{Request syntax}{
\preformatted{svc$describe_routing_profile(
  InstanceId = "string",
  RoutingProfileId = "string"
)
}
}

\keyword{internal}
