% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_contact_attributes}
\alias{connect_get_contact_attributes}
\title{Retrieves the contact attributes associated with a contact}
\usage{
connect_get_contact_attributes(InstanceId, InitialContactId)
}
\arguments{
\item{InstanceId}{[required] The instance ID for the instance from which to retrieve contact
attributes.}

\item{InitialContactId}{[required] The ID for the initial contact in Amazon Connect associated with the
attributes to update.}
}
\description{
Retrieves the contact attributes associated with a contact.
}
\section{Request syntax}{
\preformatted{connect$get_contact_attributes(
  InstanceId = "string",
  InitialContactId = "string"
)
}
}

\keyword{internal}
