% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_instance_storage_configs}
\alias{connect_list_instance_storage_configs}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_list_instance_storage_configs(InstanceId, ResourceType,
  NextToken, MaxResults)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{ResourceType}{[required] A valid resource type.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximimum number of results to return per page.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Returns a paginated list of storage configs for the identified instance
and resource type.
}
\section{Request syntax}{
\preformatted{svc$list_instance_storage_configs(
  InstanceId = "string",
  ResourceType = "CHAT_TRANSCRIPTS"|"CALL_RECORDINGS"|"SCHEDULED_REPORTS"|"MEDIA_STREAMS"|"CONTACT_TRACE_RECORDS"|"AGENT_EVENTS",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
