% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_routing_profile_name}
\alias{connect_update_routing_profile_name}
\title{Updates the name and description of a routing profile}
\usage{
connect_update_routing_profile_name(InstanceId, RoutingProfileId, Name,
  Description)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{Name}{The name of the routing profile. Must not be more than 127 characters.}

\item{Description}{The description of the routing profile. Must not be more than 250
characters.}
}
\description{
Updates the name and description of a routing profile. The request
accepts the following data in JSON format. At least \code{Name} or
\code{Description} must be provided.
}
\section{Request syntax}{
\preformatted{svc$update_routing_profile_name(
  InstanceId = "string",
  RoutingProfileId = "string",
  Name = "string",
  Description = "string"
)
}
}

\keyword{internal}
