% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_user_hierarchy}
\alias{connect_update_user_hierarchy}
\title{Assigns the specified hierarchy group to the specified user}
\usage{
connect_update_user_hierarchy(HierarchyGroupId, UserId, InstanceId)
}
\arguments{
\item{HierarchyGroupId}{The identifier of the hierarchy group.}

\item{UserId}{[required] The identifier of the user account.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\description{
Assigns the specified hierarchy group to the specified user.
}
\section{Request syntax}{
\preformatted{svc$update_user_hierarchy(
  HierarchyGroupId = "string",
  UserId = "string",
  InstanceId = "string"
)
}
}

\keyword{internal}
