% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_create_participant_connection}
\alias{connectparticipant_create_participant_connection}
\title{Creates the participant's connection}
\usage{
connectparticipant_create_participant_connection(Type, ParticipantToken)
}
\arguments{
\item{Type}{[required] Type of connection information required.}

\item{ParticipantToken}{[required] Participant Token as obtained from
\href{https://docs.aws.amazon.com/connect/latest/APIReference/API_StartChatContactResponse.html}{StartChatContact}
API response.}
}
\description{
Creates the participant\'s connection. Note that ParticipantToken is
used for invoking this API instead of ConnectionToken.
}
\details{
The participant token is valid for the lifetime of the participant --
until the they are part of a contact.

The response URL for \code{WEBSOCKET} Type has a connect expiry timeout of
100s. Clients must manually connect to the returned websocket URL and
subscribe to the desired topic.

For chat, you need to publish the following on the established websocket
connection:

\verb{\\\{"topic":"aws/subscribe","content":\\\{"topics":\\\["aws/chat"\\\]\\\}\\\}}

Upon websocket URL expiry, as specified in the response ConnectionExpiry
parameter, clients need to call this API again to obtain a new websocket
URL and perform the same steps as before.
}
\section{Request syntax}{
\preformatted{svc$create_participant_connection(
  Type = list(
    "WEBSOCKET"|"CONNECTION_CREDENTIALS"
  ),
  ParticipantToken = "string"
)
}
}

\keyword{internal}
