% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_get_transcript}
\alias{connectparticipant_get_transcript}
\title{Retrieves a transcript of the session, including details about any
attachments}
\usage{
connectparticipant_get_transcript(ContactId, MaxResults, NextToken,
  ScanDirection, SortOrder, StartPosition, ConnectionToken)
}
\arguments{
\item{ContactId}{The contactId from the current contact chain for which transcript is
needed.}

\item{MaxResults}{The maximum number of results to return in the page. Default: 10.}

\item{NextToken}{The pagination token. Use the value returned previously in the next
subsequent request to retrieve the next set of results.}

\item{ScanDirection}{The direction from StartPosition from which to retrieve message.
Default: BACKWARD when no StartPosition is provided, FORWARD with
StartPosition.}

\item{SortOrder}{The sort order for the records. Default: DESCENDING.}

\item{StartPosition}{A filtering option for where to start.}

\item{ConnectionToken}{[required] The authentication token associated with the participant's connection.}
}
\description{
Retrieves a transcript of the session, including details about any
attachments. Note that ConnectionToken is used for invoking this API
instead of ParticipantToken.
}
\details{
The Amazon Connect Participant Service APIs do not use \href{https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html}{Signature Version 4 authentication}.
}
\section{Request syntax}{
\preformatted{svc$get_transcript(
  ContactId = "string",
  MaxResults = 123,
  NextToken = "string",
  ScanDirection = "FORWARD"|"BACKWARD",
  SortOrder = "DESCENDING"|"ASCENDING",
  StartPosition = list(
    Id = "string",
    AbsoluteTime = "string",
    MostRecent = 123
  ),
  ConnectionToken = "string"
)
}
}

\keyword{internal}
