% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_send_event}
\alias{connectparticipant_send_event}
\title{Sends an event}
\usage{
connectparticipant_send_event(ContentType, Content, ClientToken,
  ConnectionToken)
}
\arguments{
\item{ContentType}{[required] The content type of the request. Supported types are:
\itemize{
\item application/vnd.amazonaws.connect.event.typing
\item application/vnd.amazonaws.connect.event.connection.acknowledged
}}

\item{Content}{The content of the event to be sent (for example, message text). This is
not yet supported.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{ConnectionToken}{[required] The authentication token associated with the participant's connection.}
}
\description{
Sends an event. Note that ConnectionToken is used for invoking this API
instead of ParticipantToken.

The Amazon Connect Participant Service APIs do not use \href{https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html}{Signature Version 4 authentication}.
}
\section{Request syntax}{
\preformatted{svc$send_event(
  ContentType = "string",
  Content = "string",
  ClientToken = "string",
  ConnectionToken = "string"
)
}
}

\keyword{internal}
