% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_send_message}
\alias{connectparticipant_send_message}
\title{Sends a message}
\usage{
connectparticipant_send_message(ContentType, Content, ClientToken,
  ConnectionToken)
}
\arguments{
\item{ContentType}{[required] The type of the content. Supported types are text/plain.}

\item{Content}{[required] The content of the message.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{ConnectionToken}{[required] The authentication token associated with the connection.}
}
\description{
Sends a message. Note that ConnectionToken is used for invoking this API
instead of ParticipantToken.

The Amazon Connect Participant Service APIs do not use \href{https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html}{Signature Version 4 authentication}.
}
\section{Request syntax}{
\preformatted{svc$send_message(
  ContentType = "string",
  Content = "string",
  ClientToken = "string",
  ConnectionToken = "string"
)
}
}

\keyword{internal}
