% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_describe_report_definitions}
\alias{costandusagereportservice_describe_report_definitions}
\title{Lists the AWS Cost and Usage reports available to this account}
\usage{
costandusagereportservice_describe_report_definitions(MaxResults,
  NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Lists the AWS Cost and Usage reports available to this account.
}
\section{Request syntax}{
\preformatted{svc$describe_report_definitions(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\examples{
# The following example lists the AWS Cost and Usage reports for the
# account.
\dontrun{svc$describe_report_definitions(
  MaxResults = 5L
)}

}
\keyword{internal}
