% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_modify_report_definition}
\alias{costandusagereportservice_modify_report_definition}
\title{Allows you to programatically update your report preferences}
\usage{
costandusagereportservice_modify_report_definition(ReportName,
  ReportDefinition)
}
\arguments{
\item{ReportName}{[required]}

\item{ReportDefinition}{[required]}
}
\description{
Allows you to programatically update your report preferences.
}
\section{Request syntax}{
\preformatted{svc$modify_report_definition(
  ReportName = "string",
  ReportDefinition = list(
    ReportName = "string",
    TimeUnit = "HOURLY"|"DAILY",
    Format = "textORcsv"|"Parquet",
    Compression = "ZIP"|"GZIP"|"Parquet",
    AdditionalSchemaElements = list(
      "RESOURCES"
    ),
    S3Bucket = "string",
    S3Prefix = "string",
    S3Region = "us-east-1"|"us-west-1"|"us-west-2"|"eu-central-1"|"eu-west-1"|"ap-southeast-1"|"ap-southeast-2"|"ap-northeast-1"|"eu-north-1"|"ap-northeast-3"|"ap-east-1",
    AdditionalArtifacts = list(
      "REDSHIFT"|"QUICKSIGHT"|"ATHENA"
    ),
    RefreshClosedReports = TRUE|FALSE,
    ReportVersioning = "CREATE_NEW_REPORT"|"OVERWRITE_REPORT"
  )
)
}
}

\keyword{internal}
