% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{create_strain_standard}
\alias{create_strain_standard}
\title{Create new strain standards}
\usage{
create_strain_standard(paw.features, strain)
}
\arguments{
\item{paw.features}{list of extracted paw features}

\item{strain}{character string or a vector of strain names, each of which
is a character string, matching in length and order the list of
paw features.}
}
\value{
list of pre-peak and post-peak strain standards, indexed by strain
}
\description{
Creates a new reference centering and scaling for the paw features of the
given strain or strains.
}
