% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_add_profile_key}
\alias{customerprofiles_add_profile_key}
\title{Associates a new key value with a specific profile, such as a Contact
Trace Record (CTR) ContactId}
\usage{
customerprofiles_add_profile_key(ProfileId, KeyName, Values, DomainName)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of a customer profile.}

\item{KeyName}{[required] A searchable identifier of a customer profile.}

\item{Values}{[required] A list of key values.}

\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Associates a new key value with a specific profile, such as a Contact
Trace Record (CTR) ContactId.

A profile object can have a single unique key and any number of
additional keys that can be used to identify the profile that it belongs
to.
}
\section{Request syntax}{
\preformatted{svc$add_profile_key(
  ProfileId = "string",
  KeyName = "string",
  Values = list(
    "string"
  ),
  DomainName = "string"
)
}
}

\keyword{internal}
