% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_create_domain}
\alias{customerprofiles_create_domain}
\title{Creates a domain, which is a container for all customer data, such as
customer profile attributes, object types, profile keys, and encryption
keys}
\usage{
customerprofiles_create_domain(DomainName, DefaultExpirationDays,
  DefaultEncryptionKey, DeadLetterQueueUrl, Tags)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{DefaultExpirationDays}{[required] The default number of days until the data within the domain expires.}

\item{DefaultEncryptionKey}{The default encryption key, which is an AWS managed key, is used when no
specific type of encryption key is specified. It is used to encrypt all
data before it is placed in permanent or semi-permanent storage.}

\item{DeadLetterQueueUrl}{The URL of the SQS dead letter queue, which is used for reporting errors
associated with ingesting data from third party applications. You must
set up a policy on the DeadLetterQueue for the SendMessage operation to
enable Amazon Connect Customer Profiles to send messages to the
DeadLetterQueue.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Creates a domain, which is a container for all customer data, such as
customer profile attributes, object types, profile keys, and encryption
keys. You can create multiple domains, and each domain can have multiple
third-party integrations.
}
\details{
Each Amazon Connect instance can be associated with only one domain.
Multiple Amazon Connect instances can be associated with one domain.
}
\section{Request syntax}{
\preformatted{svc$create_domain(
  DomainName = "string",
  DefaultExpirationDays = 123,
  DefaultEncryptionKey = "string",
  DeadLetterQueueUrl = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
