% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_domain}
\alias{customerprofiles_delete_domain}
\title{Deletes a specific domain and all of its customer data, such as customer
profile attributes and their related objects}
\usage{
customerprofiles_delete_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Deletes a specific domain and all of its customer data, such as customer
profile attributes and their related objects.
}
\section{Request syntax}{
\preformatted{svc$delete_domain(
  DomainName = "string"
)
}
}

\keyword{internal}
