% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_integration}
\alias{customerprofiles_delete_integration}
\title{Removes an integration from a specific domain}
\usage{
customerprofiles_delete_integration(DomainName, Uri)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{Uri}{The URI of the S3 bucket or any other type of data source.}
}
\description{
Removes an integration from a specific domain.
}
\section{Request syntax}{
\preformatted{svc$delete_integration(
  DomainName = "string",
  Uri = "string"
)
}
}

\keyword{internal}
