% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_profile_key}
\alias{customerprofiles_delete_profile_key}
\title{Removes a searchable key from a customer profile}
\usage{
customerprofiles_delete_profile_key(ProfileId, KeyName, Values,
  DomainName)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of a customer profile.}

\item{KeyName}{[required] A searchable identifier of a customer profile.}

\item{Values}{[required] A list of key values.}

\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Removes a searchable key from a customer profile.
}
\section{Request syntax}{
\preformatted{svc$delete_profile_key(
  ProfileId = "string",
  KeyName = "string",
  Values = list(
    "string"
  ),
  DomainName = "string"
)
}
}

\keyword{internal}
