% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_profile_object_type}
\alias{customerprofiles_delete_profile_object_type}
\title{Removes a ProfileObjectType from a specific domain as well as removes
all the ProfileObjects of that type}
\usage{
customerprofiles_delete_profile_object_type(DomainName, ObjectTypeName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The name of the profile object type.}
}
\description{
Removes a ProfileObjectType from a specific domain as well as removes
all the ProfileObjects of that type. It also disables integrations from
this specific ProfileObjectType. In addition, it scrubs all of the
fields of the standard profile that were populated from this
ProfileObjectType.
}
\section{Request syntax}{
\preformatted{svc$delete_profile_object_type(
  DomainName = "string",
  ObjectTypeName = "string"
)
}
}

\keyword{internal}
