% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_domain}
\alias{customerprofiles_get_domain}
\title{Returns information about a specific domain}
\usage{
customerprofiles_get_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] A unique name for the domain.}
}
\description{
Returns information about a specific domain.
}
\section{Request syntax}{
\preformatted{svc$get_domain(
  DomainName = "string"
)
}
}

\keyword{internal}
