% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_integration}
\alias{customerprofiles_get_integration}
\title{Returns an integration for a domain}
\usage{
customerprofiles_get_integration(DomainName, Uri)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{Uri}{The URI of the S3 bucket or any other type of data source.}
}
\description{
Returns an integration for a domain.
}
\section{Request syntax}{
\preformatted{svc$get_integration(
  DomainName = "string",
  Uri = "string"
)
}
}

\keyword{internal}
