% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_profile_object_type}
\alias{customerprofiles_get_profile_object_type}
\title{Returns the object types for a specific domain}
\usage{
customerprofiles_get_profile_object_type(DomainName, ObjectTypeName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The name of the profile object type.}
}
\description{
Returns the object types for a specific domain.
}
\section{Request syntax}{
\preformatted{svc$get_profile_object_type(
  DomainName = "string",
  ObjectTypeName = "string"
)
}
}

\keyword{internal}
