% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_profile_object_type_template}
\alias{customerprofiles_get_profile_object_type_template}
\title{Returns the template information for a specific object type}
\usage{
customerprofiles_get_profile_object_type_template(TemplateId)
}
\arguments{
\item{TemplateId}{[required] A unique identifier for the object template.}
}
\description{
Returns the template information for a specific object type.

A template is a predefined ProfileObjectType, such as
“Salesforce-Account” or “Salesforce-Contact.” When a user sends a
ProfileObject, using the PutProfileObject API, with an ObjectTypeName
that matches one of the TemplateIds, it uses the mappings from the
template.
}
\section{Request syntax}{
\preformatted{svc$get_profile_object_type_template(
  TemplateId = "string"
)
}
}

\keyword{internal}
