% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_integrations}
\alias{customerprofiles_list_integrations}
\title{Lists all of the integrations in your domain}
\usage{
customerprofiles_list_integrations(DomainName, NextToken, MaxResults)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{NextToken}{The pagination token from the previous ListIntegrations API call.}

\item{MaxResults}{The maximum number of objects returned per page.}
}
\description{
Lists all of the integrations in your domain.
}
\section{Request syntax}{
\preformatted{svc$list_integrations(
  DomainName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
