% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_profile_object_type_templates}
\alias{customerprofiles_list_profile_object_type_templates}
\title{Lists all of the template information for object types}
\usage{
customerprofiles_list_profile_object_type_templates(NextToken,
  MaxResults)
}
\arguments{
\item{NextToken}{The pagination token from the previous ListObjectTypeTemplates API call.}

\item{MaxResults}{The maximum number of objects returned per page.}
}
\description{
Lists all of the template information for object types.
}
\section{Request syntax}{
\preformatted{svc$list_profile_object_type_templates(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
