% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_profile_objects}
\alias{customerprofiles_list_profile_objects}
\title{Returns a list of objects associated with a profile of a given
ProfileObjectType}
\usage{
customerprofiles_list_profile_objects(NextToken, MaxResults, DomainName,
  ObjectTypeName, ProfileId)
}
\arguments{
\item{NextToken}{The pagination token from the previous call to ListProfileObjects.}

\item{MaxResults}{The maximum number of objects returned per page.}

\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The name of the profile object type.}

\item{ProfileId}{[required] The unique identifier of a customer profile.}
}
\description{
Returns a list of objects associated with a profile of a given
ProfileObjectType.
}
\section{Request syntax}{
\preformatted{svc$list_profile_objects(
  NextToken = "string",
  MaxResults = 123,
  DomainName = "string",
  ObjectTypeName = "string",
  ProfileId = "string"
)
}
}

\keyword{internal}
