% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_tags_for_resource}
\alias{customerprofiles_list_tags_for_resource}
\title{Displays the tags associated with an Amazon Connect Customer Profiles
resource}
\usage{
customerprofiles_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource for which you want to view tags.}
}
\description{
Displays the tags associated with an Amazon Connect Customer Profiles
resource. In Connect Customer Profiles, domains, profile object types,
and integrations can be tagged.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
