% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_put_integration}
\alias{customerprofiles_put_integration}
\title{Adds an integration between the service and a third-party service, which
includes Amazon AppFlow and Amazon Connect}
\usage{
customerprofiles_put_integration(DomainName, Uri, ObjectTypeName, Tags)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{Uri}{[required] The URI of the S3 bucket or any other type of data source.}

\item{ObjectTypeName}{[required] The name of the profile object type.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Adds an integration between the service and a third-party service, which
includes Amazon AppFlow and Amazon Connect.

An integration can belong to only one domain.
}
\section{Request syntax}{
\preformatted{svc$put_integration(
  DomainName = "string",
  Uri = "string",
  ObjectTypeName = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
