% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_put_profile_object}
\alias{customerprofiles_put_profile_object}
\title{Adds additional objects to customer profiles of a given ObjectType}
\usage{
customerprofiles_put_profile_object(ObjectTypeName, Object, DomainName)
}
\arguments{
\item{ObjectTypeName}{[required] The name of the profile object type.}

\item{Object}{[required] A string that is serialized from a JSON object.}

\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Adds additional objects to customer profiles of a given ObjectType.

When adding a specific profile object, like a Contact Trace Record
(CTR), an inferred profile can get created if it is not mapped to an
existing profile. The resulting profile will only have a phone number
populated in the standard ProfileObject. Any additional CTRs with the
same phone number will be mapped to the same inferred profile.

When a ProfileObject is created and if a ProfileObjectType already
exists for the ProfileObject, it will provide data to a standard profile
depending on the ProfileObjectType definition.

PutProfileObject needs an ObjectType, which can be created using
PutProfileObjectType.
}
\section{Request syntax}{
\preformatted{svc$put_profile_object(
  ObjectTypeName = "string",
  Object = "string",
  DomainName = "string"
)
}
}

\keyword{internal}
