% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_search_profiles}
\alias{customerprofiles_search_profiles}
\title{Searches for profiles within a specific domain name using name, phone
number, email address, account number, or a custom defined index}
\usage{
customerprofiles_search_profiles(NextToken, MaxResults, DomainName,
  KeyName, Values)
}
\arguments{
\item{NextToken}{The pagination token from the previous SearchProfiles API call.}

\item{MaxResults}{The maximum number of objects returned per page.}

\item{DomainName}{[required] The unique name of the domain.}

\item{KeyName}{[required] A searchable identifier of a customer profile. The predefined keys you
can use to search include: \\_account, \\_profileId, \\_fullName, \\_phone,
\\_email, \\_ctrContactId, \\_marketoLeadId, \\_salesforceAccountId,
\\_salesforceContactId, \\_zendeskUserId, \\_zendeskExternalId,
\\_serviceNowSystemId.}

\item{Values}{[required] A list of key values.}
}
\description{
Searches for profiles within a specific domain name using name, phone
number, email address, account number, or a custom defined index.
}
\section{Request syntax}{
\preformatted{svc$search_profiles(
  NextToken = "string",
  MaxResults = 123,
  DomainName = "string",
  KeyName = "string",
  Values = list(
    "string"
  )
)
}
}

\keyword{internal}
