% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_update_domain}
\alias{customerprofiles_update_domain}
\title{Updates the properties of a domain, including creating or selecting a
dead letter queue or an encryption key}
\usage{
customerprofiles_update_domain(DomainName, DefaultExpirationDays,
  DefaultEncryptionKey, DeadLetterQueueUrl, Tags)
}
\arguments{
\item{DomainName}{[required] The unique name for the domain.}

\item{DefaultExpirationDays}{The default number of days until the data within the domain expires.}

\item{DefaultEncryptionKey}{The default encryption key, which is an AWS managed key, is used when no
specific type of encryption key is specified. It is used to encrypt all
data before it is placed in permanent or semi-permanent storage. If
specified as an empty string, it will clear any existing value.}

\item{DeadLetterQueueUrl}{The URL of the SQS dead letter queue, which is used for reporting errors
associated with ingesting data from third party applications. If
specified as an empty string, it will clear any existing value. You must
set up a policy on the DeadLetterQueue for the SendMessage operation to
enable Amazon Connect Customer Profiles to send messages to the
DeadLetterQueue.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Updates the properties of a domain, including creating or selecting a
dead letter queue or an encryption key.

Once a domain is created, the name can’t be changed.
}
\section{Request syntax}{
\preformatted{svc$update_domain(
  DomainName = "string",
  DefaultExpirationDays = 123,
  DefaultEncryptionKey = "string",
  DeadLetterQueueUrl = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
