% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_update_profile}
\alias{customerprofiles_update_profile}
\title{Updates the properties of a profile}
\usage{
customerprofiles_update_profile(DomainName, ProfileId,
  AdditionalInformation, AccountNumber, PartyType, BusinessName,
  FirstName, MiddleName, LastName, BirthDate, Gender, PhoneNumber,
  MobilePhoneNumber, HomePhoneNumber, BusinessPhoneNumber, EmailAddress,
  PersonalEmailAddress, BusinessEmailAddress, Address, ShippingAddress,
  MailingAddress, BillingAddress, Attributes)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ProfileId}{[required] The unique identifier of a customer profile.}

\item{AdditionalInformation}{Any additional information relevant to the customer's profile.}

\item{AccountNumber}{A unique account number that you have given to the customer.}

\item{PartyType}{The type of profile used to describe the customer.}

\item{BusinessName}{The name of the customer’s business.}

\item{FirstName}{The customer’s first name.}

\item{MiddleName}{The customer’s middle name.}

\item{LastName}{The customer’s last name.}

\item{BirthDate}{The customer’s birth date.}

\item{Gender}{The gender with which the customer identifies.}

\item{PhoneNumber}{The customer's phone number, which has not been specified as a mobile,
home, or business number.}

\item{MobilePhoneNumber}{The customer’s mobile phone number.}

\item{HomePhoneNumber}{The customer’s home phone number.}

\item{BusinessPhoneNumber}{The customer’s business phone number.}

\item{EmailAddress}{The customer's email address, which has not been specified as a personal
or business address.}

\item{PersonalEmailAddress}{The customer’s personal email address.}

\item{BusinessEmailAddress}{The customer’s business email address.}

\item{Address}{A generic address associated with the customer that is not mailing,
shipping, or billing.}

\item{ShippingAddress}{The customer’s shipping address.}

\item{MailingAddress}{The customer’s mailing address.}

\item{BillingAddress}{The customer’s billing address.}

\item{Attributes}{A key value pair of attributes of a customer profile.}
}
\description{
Updates the properties of a profile. The ProfileId is required for
updating a customer profile.

When calling the UpdateProfile API, specifying an empty string value
means that any existing value will be removed. Not specifying a string
value means that any value already there will be kept.
}
\section{Request syntax}{
\preformatted{svc$update_profile(
  DomainName = "string",
  ProfileId = "string",
  AdditionalInformation = "string",
  AccountNumber = "string",
  PartyType = "INDIVIDUAL"|"BUSINESS"|"OTHER",
  BusinessName = "string",
  FirstName = "string",
  MiddleName = "string",
  LastName = "string",
  BirthDate = "string",
  Gender = "MALE"|"FEMALE"|"UNSPECIFIED",
  PhoneNumber = "string",
  MobilePhoneNumber = "string",
  HomePhoneNumber = "string",
  BusinessPhoneNumber = "string",
  EmailAddress = "string",
  PersonalEmailAddress = "string",
  BusinessEmailAddress = "string",
  Address = list(
    Address1 = "string",
    Address2 = "string",
    Address3 = "string",
    Address4 = "string",
    City = "string",
    County = "string",
    State = "string",
    Province = "string",
    Country = "string",
    PostalCode = "string"
  ),
  ShippingAddress = list(
    Address1 = "string",
    Address2 = "string",
    Address3 = "string",
    Address4 = "string",
    City = "string",
    County = "string",
    State = "string",
    Province = "string",
    Country = "string",
    PostalCode = "string"
  ),
  MailingAddress = list(
    Address1 = "string",
    Address2 = "string",
    Address3 = "string",
    Address4 = "string",
    City = "string",
    County = "string",
    State = "string",
    Province = "string",
    Country = "string",
    PostalCode = "string"
  ),
  BillingAddress = list(
    Address1 = "string",
    Address2 = "string",
    Address3 = "string",
    Address4 = "string",
    City = "string",
    County = "string",
    State = "string",
    Province = "string",
    Country = "string",
    PostalCode = "string"
  ),
  Attributes = list(
    "string"
  )
)
}
}

\keyword{internal}
