% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_service.R
\name{databasemigrationservice}
\alias{databasemigrationservice}
\title{AWS Database Migration Service}
\usage{
databasemigrationservice()
}
\description{
AWS Database Migration Service (AWS DMS) can migrate your data to and
from the most widely used commercial and open-source databases such as
Oracle, PostgreSQL, Microsoft SQL Server, Amazon Redshift, MariaDB,
Amazon Aurora, MySQL, and SAP Adaptive Server Enterprise (ASE). The
service supports homogeneous migrations such as Oracle to Oracle, as
well as heterogeneous migrations between different database platforms,
such as Oracle to MySQL or SQL Server to PostgreSQL.

For more information about AWS DMS, see \href{http://docs.aws.amazon.com/dms/latest/userguide/Welcome.html}{What Is AWS Database Migration Service?}
in the \emph{AWS Database Migration User Guide.}
}
\section{Example}{
\preformatted{databasemigrationservice <- paws::databasemigrationservice()
databasemigrationservice$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=databasemigrationservice_add_tags_to_resource]{add_tags_to_resource} \tab Adds metadata tags to an AWS DMS resource, including replication instance, endpoint, security group, and migration task \cr
\link[=databasemigrationservice_create_endpoint]{create_endpoint} \tab Creates an endpoint using the provided settings \cr
\link[=databasemigrationservice_create_event_subscription]{create_event_subscription} \tab Creates an AWS DMS event notification subscription \cr
\link[=databasemigrationservice_create_replication_instance]{create_replication_instance} \tab Creates the replication instance using the specified parameters \cr
\link[=databasemigrationservice_create_replication_subnet_group]{create_replication_subnet_group} \tab Creates a replication subnet group given a list of the subnet IDs in a VPC \cr
\link[=databasemigrationservice_create_replication_task]{create_replication_task} \tab Creates a replication task using the specified parameters \cr
\link[=databasemigrationservice_delete_certificate]{delete_certificate} \tab Deletes the specified certificate \cr
\link[=databasemigrationservice_delete_endpoint]{delete_endpoint} \tab Deletes the specified endpoint \cr
\link[=databasemigrationservice_delete_event_subscription]{delete_event_subscription} \tab Deletes an AWS DMS event subscription \cr
\link[=databasemigrationservice_delete_replication_instance]{delete_replication_instance} \tab Deletes the specified replication instance \cr
\link[=databasemigrationservice_delete_replication_subnet_group]{delete_replication_subnet_group} \tab Deletes a subnet group \cr
\link[=databasemigrationservice_delete_replication_task]{delete_replication_task} \tab Deletes the specified replication task \cr
\link[=databasemigrationservice_describe_account_attributes]{describe_account_attributes} \tab Lists all of the AWS DMS attributes for a customer account \cr
\link[=databasemigrationservice_describe_certificates]{describe_certificates} \tab Provides a description of the certificate \cr
\link[=databasemigrationservice_describe_connections]{describe_connections} \tab Describes the status of the connections that have been made between the replication instance and an endpoint \cr
\link[=databasemigrationservice_describe_endpoint_types]{describe_endpoint_types} \tab Returns information about the type of endpoints available \cr
\link[=databasemigrationservice_describe_endpoints]{describe_endpoints} \tab Returns information about the endpoints for your account in the current region \cr
\link[=databasemigrationservice_describe_event_categories]{describe_event_categories} \tab Lists categories for all event source types, or, if specified, for a specified source type \cr
\link[=databasemigrationservice_describe_event_subscriptions]{describe_event_subscriptions} \tab Lists all the event subscriptions for a customer account \cr
\link[=databasemigrationservice_describe_events]{describe_events} \tab Lists events for a given source identifier and source type \cr
\link[=databasemigrationservice_describe_orderable_replication_instances]{describe_orderable_replication_instances} \tab Returns information about the replication instance types that can be created in the specified region \cr
\link[=databasemigrationservice_describe_refresh_schemas_status]{describe_refresh_schemas_status} \tab Returns the status of the RefreshSchemas operation \cr
\link[=databasemigrationservice_describe_replication_instance_task_logs]{describe_replication_instance_task_logs} \tab Returns information about the task logs for the specified task \cr
\link[=databasemigrationservice_describe_replication_instances]{describe_replication_instances} \tab Returns information about replication instances for your account in the current region \cr
\link[=databasemigrationservice_describe_replication_subnet_groups]{describe_replication_subnet_groups} \tab Returns information about the replication subnet groups \cr
\link[=databasemigrationservice_describe_replication_task_assessment_results]{describe_replication_task_assessment_results} \tab Returns the task assessment results from Amazon S3 \cr
\link[=databasemigrationservice_describe_replication_tasks]{describe_replication_tasks} \tab Returns information about replication tasks for your account in the current region \cr
\link[=databasemigrationservice_describe_schemas]{describe_schemas} \tab Returns information about the schema for the specified endpoint \cr
\link[=databasemigrationservice_describe_table_statistics]{describe_table_statistics} \tab Returns table statistics on the database migration task, including table name, rows inserted, rows updated, and rows deleted\cr
\link[=databasemigrationservice_import_certificate]{import_certificate} \tab Uploads the specified certificate \cr
\link[=databasemigrationservice_list_tags_for_resource]{list_tags_for_resource} \tab Lists all tags for an AWS DMS resource \cr
\link[=databasemigrationservice_modify_endpoint]{modify_endpoint} \tab Modifies the specified endpoint \cr
\link[=databasemigrationservice_modify_event_subscription]{modify_event_subscription} \tab Modifies an existing AWS DMS event notification subscription \cr
\link[=databasemigrationservice_modify_replication_instance]{modify_replication_instance} \tab Modifies the replication instance to apply new settings \cr
\link[=databasemigrationservice_modify_replication_subnet_group]{modify_replication_subnet_group} \tab Modifies the settings for the specified replication subnet group \cr
\link[=databasemigrationservice_modify_replication_task]{modify_replication_task} \tab Modifies the specified replication task \cr
\link[=databasemigrationservice_reboot_replication_instance]{reboot_replication_instance} \tab Reboots a replication instance \cr
\link[=databasemigrationservice_refresh_schemas]{refresh_schemas} \tab Populates the schema for the specified endpoint \cr
\link[=databasemigrationservice_reload_tables]{reload_tables} \tab Reloads the target database table with the source data \cr
\link[=databasemigrationservice_remove_tags_from_resource]{remove_tags_from_resource} \tab Removes metadata tags from a DMS resource \cr
\link[=databasemigrationservice_start_replication_task]{start_replication_task} \tab Starts the replication task \cr
\link[=databasemigrationservice_start_replication_task_assessment]{start_replication_task_assessment} \tab Starts the replication task assessment for unsupported data types in the source database \cr
\link[=databasemigrationservice_stop_replication_task]{stop_replication_task} \tab Stops the replication task \cr
\link[=databasemigrationservice_test_connection]{test_connection} \tab Tests the connection between the replication instance and the endpoint
}
}

