% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_add_tags_to_resource}
\alias{databasemigrationservice_add_tags_to_resource}
\title{Adds metadata tags to an AWS DMS resource, including replication
instance, endpoint, security group, and migration task}
\usage{
databasemigrationservice_add_tags_to_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the AWS DMS resource the tag is to be
added to. AWS DMS resources include a replication instance, endpoint,
and a replication task.}

\item{Tags}{[required] The tag to be assigned to the DMS resource.}
}
\description{
Adds metadata tags to an AWS DMS resource, including replication
instance, endpoint, security group, and migration task. These tags can
also be used with cost allocation reporting to track cost associated
with DMS resources, or used in a Condition statement in an IAM policy
for DMS.
}
\section{Request syntax}{
\preformatted{databasemigrationservice$add_tags_to_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\examples{
# Adds metadata tags to an AWS DMS resource, including replication
# instance, endpoint, security group, and migration task. These tags can
# also be used with cost allocation reporting to track cost associated
# with AWS DMS resources, or used in a Condition statement in an IAM
# policy for AWS DMS.
\donttest{databasemigrationservice$add_tags_to_resource(
  ResourceArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7E",
  Tags = list(
    list(
      Key = "Acount",
      Value = "1633456"
    )
  )
)}

}
\keyword{internal}
