% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_cancel_replication_task_assessment_run}
\alias{databasemigrationservice_cancel_replication_task_assessment_run}
\title{Cancels a single premigration assessment run}
\usage{
databasemigrationservice_cancel_replication_task_assessment_run(
  ReplicationTaskAssessmentRunArn)
}
\arguments{
\item{ReplicationTaskAssessmentRunArn}{[required] Amazon Resource Name (ARN) of the premigration assessment run to be
canceled.}
}
\description{
Cancels a single premigration assessment run.

This operation prevents any individual assessments from running if they
haven't started running. It also attempts to cancel any individual
assessments that are currently running.
}
\section{Request syntax}{
\preformatted{svc$cancel_replication_task_assessment_run(
  ReplicationTaskAssessmentRunArn = "string"
)
}
}

\keyword{internal}
