% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_create_replication_instance}
\alias{databasemigrationservice_create_replication_instance}
\title{Creates the replication instance using the specified parameters}
\usage{
databasemigrationservice_create_replication_instance(
  ReplicationInstanceIdentifier, AllocatedStorage,
  ReplicationInstanceClass, VpcSecurityGroupIds, AvailabilityZone,
  ReplicationSubnetGroupIdentifier, PreferredMaintenanceWindow, MultiAZ,
  EngineVersion, AutoMinorVersionUpgrade, Tags, KmsKeyId,
  PubliclyAccessible, DnsNameServers)
}
\arguments{
\item{ReplicationInstanceIdentifier}{[required] The replication instance identifier. This parameter is stored as a
lowercase string.

Constraints:
\itemize{
\item Must contain from 1 to 63 alphanumeric characters or hyphens.
\item First character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
}

Example: \code{myrepinstance}}

\item{AllocatedStorage}{The amount of storage (in gigabytes) to be initially allocated for the
replication instance.}

\item{ReplicationInstanceClass}{[required] The compute and memory capacity of the replication instance as specified
by the replication instance class.

Valid Values:
\code{dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge }}

\item{VpcSecurityGroupIds}{Specifies the VPC security group to be used with the replication
instance. The VPC security group must work with the VPC containing the
replication instance.}

\item{AvailabilityZone}{The AWS Availability Zone where the replication instance will be
created. The default value is a random, system-chosen Availability Zone
in the endpoint\'s AWS Region, for example: \verb{us-east-1d}}

\item{ReplicationSubnetGroupIdentifier}{A subnet group to associate with the replication instance.}

\item{PreferredMaintenanceWindow}{The weekly time range during which system maintenance can occur, in
Universal Coordinated Time (UTC).

Format: \code{ddd:hh24:mi-ddd:hh24:mi}

Default: A 30-minute window selected at random from an 8-hour block of
time per AWS Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.}

\item{MultiAZ}{Specifies whether the replication instance is a Multi-AZ deployment. You
cannot set the \code{AvailabilityZone} parameter if the Multi-AZ parameter is
set to \code{true}.}

\item{EngineVersion}{The engine version number of the replication instance.}

\item{AutoMinorVersionUpgrade}{Indicates whether minor engine upgrades will be applied automatically to
the replication instance during the maintenance window. This parameter
defaults to \code{true}.

Default: \code{true}}

\item{Tags}{One or more tags to be assigned to the replication instance.}

\item{KmsKeyId}{An AWS KMS key identifier that is used to encrypt the data on the
replication instance.

If you don\'t specify a value for the \code{KmsKeyId} parameter, then AWS DMS
uses your default encryption key.

AWS KMS creates the default encryption key for your AWS account. Your
AWS account has a different default encryption key for each AWS Region.}

\item{PubliclyAccessible}{Specifies the accessibility options for the replication instance. A
value of \code{true} represents an instance with a public IP address. A value
of \code{false} represents an instance with a private IP address. The default
value is \code{true}.}

\item{DnsNameServers}{A list of DNS name servers supported for the replication instance.}
}
\description{
Creates the replication instance using the specified parameters.
}
\details{
AWS DMS requires that your account have certain roles with appropriate
permissions before you can create a replication instance. For
information on the required roles, see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Security.APIRole.html}{Creating the IAM Roles to Use With the AWS CLI and AWS DMS API}.
For information on the required permissions, see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Security.IAMPermissions.html}{IAM Permissions Needed to Use AWS DMS}.
}
\section{Request syntax}{
\preformatted{svc$create_replication_instance(
  ReplicationInstanceIdentifier = "string",
  AllocatedStorage = 123,
  ReplicationInstanceClass = "string",
  VpcSecurityGroupIds = list(
    "string"
  ),
  AvailabilityZone = "string",
  ReplicationSubnetGroupIdentifier = "string",
  PreferredMaintenanceWindow = "string",
  MultiAZ = TRUE|FALSE,
  EngineVersion = "string",
  AutoMinorVersionUpgrade = TRUE|FALSE,
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  KmsKeyId = "string",
  PubliclyAccessible = TRUE|FALSE,
  DnsNameServers = "string"
)
}
}

\examples{
\dontrun{
# Creates the replication instance using the specified parameters.
svc$create_replication_instance(
  AllocatedStorage = 123L,
  AutoMinorVersionUpgrade = TRUE,
  AvailabilityZone = "",
  EngineVersion = "",
  KmsKeyId = "",
  MultiAZ = TRUE,
  PreferredMaintenanceWindow = "",
  PubliclyAccessible = TRUE,
  ReplicationInstanceClass = "",
  ReplicationInstanceIdentifier = "",
  ReplicationSubnetGroupIdentifier = "",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  VpcSecurityGroupIds = list()
)
}

}
\keyword{internal}
