% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_create_replication_subnet_group}
\alias{databasemigrationservice_create_replication_subnet_group}
\title{Creates a replication subnet group given a list of the subnet IDs in a
VPC}
\usage{
databasemigrationservice_create_replication_subnet_group(
  ReplicationSubnetGroupIdentifier, ReplicationSubnetGroupDescription,
  SubnetIds, Tags)
}
\arguments{
\item{ReplicationSubnetGroupIdentifier}{[required] The name for the replication subnet group. This value is stored as a
lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters,
periods, spaces, underscores, or hyphens. Must not be \"default\".

Example: \code{mySubnetgroup}}

\item{ReplicationSubnetGroupDescription}{[required] The description for the subnet group.}

\item{SubnetIds}{[required] One or more subnet IDs to be assigned to the subnet group.}

\item{Tags}{One or more tags to be assigned to the subnet group.}
}
\description{
Creates a replication subnet group given a list of the subnet IDs in a
VPC.
}
\section{Request syntax}{
\preformatted{svc$create_replication_subnet_group(
  ReplicationSubnetGroupIdentifier = "string",
  ReplicationSubnetGroupDescription = "string",
  SubnetIds = list(
    "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\examples{
# Creates a replication subnet group given a list of the subnet IDs in a
# VPC.
\dontrun{svc$create_replication_subnet_group(
  ReplicationSubnetGroupDescription = "US West subnet group",
  ReplicationSubnetGroupIdentifier = "us-west-2ab-vpc-215ds366",
  SubnetIds = list(
    "subnet-e145356n",
    "subnet-58f79200"
  ),
  Tags = list(
    list(
      Key = "Acount",
      Value = "145235"
    )
  )
)}

}
\keyword{internal}
