% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_replication_task_assessment_runs}
\alias{databasemigrationservice_describe_replication_task_assessment_runs}
\title{Returns a paginated list of premigration assessment runs based on filter
settings}
\usage{
databasemigrationservice_describe_replication_task_assessment_runs(
  Filters, MaxRecords, Marker)
}
\arguments{
\item{Filters}{Filters applied to the premigration assessment runs described in the
form of key-value pairs.

Valid filter names: \code{replication-task-assessment-run-arn},
\code{replication-task-arn}, \code{replication-instance-arn}, \code{status}}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Returns a paginated list of premigration assessment runs based on filter
settings.

These filter settings can specify a combination of premigration
assessment runs, migration tasks, replication instances, and assessment
run status values.

This operation doesn't return information about individual assessments.
For this information, see the
\code{\link[=databasemigrationservice_describe_replication_task_individual_assessments]{describe_replication_task_individual_assessments}}
operation.
}
\section{Request syntax}{
\preformatted{svc$describe_replication_task_assessment_runs(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}
}

\keyword{internal}
