% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_table_statistics}
\alias{databasemigrationservice_describe_table_statistics}
\title{Returns table statistics on the database migration task, including table
name, rows inserted, rows updated, and rows deleted}
\usage{
databasemigrationservice_describe_table_statistics(ReplicationTaskArn,
  MaxRecords, Marker, Filters)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name (ARN) of the replication task.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 500.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{Filters}{Filters applied to the describe table statistics action.

Valid filter names: schema-name | table-name | table-state

A combination of filters creates an AND condition where each record
matches all specified filters.}
}
\description{
Returns table statistics on the database migration task, including table
name, rows inserted, rows updated, and rows deleted.
}
\details{
Note that the "last updated" column the DMS console only indicates the
time that AWS DMS last updated the table statistics record for a table.
It does not indicate the time of the last update to the table.
}
\section{Request syntax}{
\preformatted{svc$describe_table_statistics(
  ReplicationTaskArn = "string",
  MaxRecords = 123,
  Marker = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}
}

\examples{
\dontrun{
# Returns table statistics on the database migration task, including table
# name, rows inserted, rows updated, and rows deleted.
svc$describe_table_statistics(
  Marker = "",
  MaxRecords = 123L,
  ReplicationTaskArn = ""
)
}

}
\keyword{internal}
