% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_modify_replication_instance}
\alias{databasemigrationservice_modify_replication_instance}
\title{Modifies the replication instance to apply new settings}
\usage{
databasemigrationservice_modify_replication_instance(
  ReplicationInstanceArn, AllocatedStorage, ApplyImmediately,
  ReplicationInstanceClass, VpcSecurityGroupIds,
  PreferredMaintenanceWindow, MultiAZ, EngineVersion,
  AllowMajorVersionUpgrade, AutoMinorVersionUpgrade,
  ReplicationInstanceIdentifier)
}
\arguments{
\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance.}

\item{AllocatedStorage}{The amount of storage (in gigabytes) to be allocated for the replication
instance.}

\item{ApplyImmediately}{Indicates whether the changes should be applied immediately or during
the next maintenance window.}

\item{ReplicationInstanceClass}{The compute and memory capacity of the replication instance.

Valid Values:
\code{dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge }}

\item{VpcSecurityGroupIds}{Specifies the VPC security group to be used with the replication
instance. The VPC security group must work with the VPC containing the
replication instance.}

\item{PreferredMaintenanceWindow}{The weekly time range (in UTC) during which system maintenance can
occur, which might result in an outage. Changing this parameter does not
result in an outage, except in the following situation, and the change
is asynchronously applied as soon as possible. If moving this window to
the current time, there must be at least 30 minutes between the current
time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon \| Tue \| Wed \| Thu \| Fri \| Sat \| Sun

Constraints: Must be at least 30 minutes}

\item{MultiAZ}{Specifies whether the replication instance is a Multi-AZ deployment. You
cannot set the \code{AvailabilityZone} parameter if the Multi-AZ parameter is
set to \code{true}.}

\item{EngineVersion}{The engine version number of the replication instance.}

\item{AllowMajorVersionUpgrade}{Indicates that major version upgrades are allowed. Changing this
parameter does not result in an outage, and the change is asynchronously
applied as soon as possible.

This parameter must be set to \code{true} when specifying a value for the
\code{EngineVersion} parameter that is a different major version than the
replication instance's current version.}

\item{AutoMinorVersionUpgrade}{Indicates that minor version upgrades will be applied automatically to
the replication instance during the maintenance window. Changing this
parameter does not result in an outage except in the following case and
the change is asynchronously applied as soon as possible. An outage will
result if this parameter is set to \code{true} during the maintenance window,
and a newer minor version is available, and AWS DMS has enabled auto
patching for that engine version.}

\item{ReplicationInstanceIdentifier}{The replication instance identifier. This parameter is stored as a
lowercase string.}
}
\description{
Modifies the replication instance to apply new settings. You can change
one or more parameters by specifying these parameters and the new values
in the request.
}
\details{
Some settings are applied during the maintenance window.
}
\section{Request syntax}{
\preformatted{svc$modify_replication_instance(
  ReplicationInstanceArn = "string",
  AllocatedStorage = 123,
  ApplyImmediately = TRUE|FALSE,
  ReplicationInstanceClass = "string",
  VpcSecurityGroupIds = list(
    "string"
  ),
  PreferredMaintenanceWindow = "string",
  MultiAZ = TRUE|FALSE,
  EngineVersion = "string",
  AllowMajorVersionUpgrade = TRUE|FALSE,
  AutoMinorVersionUpgrade = TRUE|FALSE,
  ReplicationInstanceIdentifier = "string"
)
}
}

\examples{
# Modifies the replication instance to apply new settings. You can change
# one or more parameters by specifying these parameters and the new values
# in the request. Some settings are applied during the maintenance window.
\donttest{svc$modify_replication_instance(
  AllocatedStorage = 123L,
  AllowMajorVersionUpgrade = TRUE,
  ApplyImmediately = TRUE,
  AutoMinorVersionUpgrade = TRUE,
  EngineVersion = "1.5.0",
  MultiAZ = TRUE,
  PreferredMaintenanceWindow = "sun:06:00-sun:14:00",
  ReplicationInstanceArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJ...",
  ReplicationInstanceClass = "dms.t2.micro",
  ReplicationInstanceIdentifier = "test-rep-1",
  VpcSecurityGroupIds = list()
)}

}
\keyword{internal}
