% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_move_replication_task}
\alias{databasemigrationservice_move_replication_task}
\title{Moves a replication task from its current replication instance to a
different target replication instance using the specified parameters}
\usage{
databasemigrationservice_move_replication_task(ReplicationTaskArn,
  TargetReplicationInstanceArn)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name (ARN) of the task that you want to move.}

\item{TargetReplicationInstanceArn}{[required] The ARN of the replication instance where you want to move the task to.}
}
\description{
Moves a replication task from its current replication instance to a
different target replication instance using the specified parameters.
The target replication instance must be created with the same or later
AWS DMS version as the current replication instance.
}
\section{Request syntax}{
\preformatted{svc$move_replication_task(
  ReplicationTaskArn = "string",
  TargetReplicationInstanceArn = "string"
)
}
}

\keyword{internal}
