% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_remove_tags_from_resource}
\alias{databasemigrationservice_remove_tags_from_resource}
\title{Removes metadata tags from a DMS resource}
\usage{
databasemigrationservice_remove_tags_from_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] An AWS DMS resource from which you want to remove tag(s). The value for
this parameter is an Amazon Resource Name (ARN).}

\item{TagKeys}{[required] The tag key (name) of the tag to be removed.}
}
\description{
Removes metadata tags from a DMS resource.
}
\section{Request syntax}{
\preformatted{svc$remove_tags_from_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\section{Examples}{
\preformatted{# Removes metadata tags from an AWS DMS resource.
svc$remove_tags_from_resource(
  ResourceArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7E",
  TagKeys = list()
)
}
}

\keyword{internal}
