% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_start_replication_task}
\alias{databasemigrationservice_start_replication_task}
\title{Starts the replication task}
\usage{
databasemigrationservice_start_replication_task(ReplicationTaskArn,
  StartReplicationTaskType, CdcStartTime, CdcStartPosition,
  CdcStopPosition)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name (ARN) of the replication task to be started.}

\item{StartReplicationTaskType}{[required] The type of replication task.}

\item{CdcStartTime}{Indicates the start time for a change data capture (CDC) operation. Use
either CdcStartTime or CdcStartPosition to specify when you want a CDC
operation to start. Specifying both values results in an error.

Timestamp Example: \--cdc-start-time "2018-03-08T12:12:12"}

\item{CdcStartPosition}{Indicates when you want a change data capture (CDC) operation to start.
Use either CdcStartPosition or CdcStartTime to specify when you want a
CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: \--cdc-start-position "2018-03-08T12:12:12"

Checkpoint Example: \--cdc-start-position
"checkpoint:V1\#27\#mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:1876\#0\#0\#*\#0\#93"

LSN Example: \--cdc-start-position "mysql-bin-changelog.000024:373"}

\item{CdcStopPosition}{Indicates when you want a change data capture (CDC) operation to stop.
The value can be either server time or commit time.

Server time example: \--cdc-stop-position
"server\_time:3018-02-09T12:12:12"

Commit time example: \--cdc-stop-position "commit\_time:
3018-02-09T12:12:12 "}
}
\description{
Starts the replication task.
}
\details{
For more information about AWS DMS tasks, see \href{http://docs.aws.amazon.com/dms/latest/userguide/CHAP_Tasks.html}{Working with Migration Tasks}
in the \emph{AWS Database Migration Service User Guide.}
}
\section{Request syntax}{
\preformatted{databasemigrationservice$start_replication_task(
  ReplicationTaskArn = "string",
  StartReplicationTaskType = "start-replication"|"resume-processing"|"reload-target",
  CdcStartTime = as.POSIXct(
    "2015-01-01"
  ),
  CdcStartPosition = "string",
  CdcStopPosition = "string"
)
}
}

\examples{
# Starts the replication task.
\donttest{databasemigrationservice$start_replication_task(
  CdcStartTime = "2016-12-14T13:33:20Z",
  ReplicationTaskArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJQ",
  StartReplicationTaskType = "start-replication"
)}

}
\keyword{internal}
