% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_stop_replication_task}
\alias{databasemigrationservice_stop_replication_task}
\title{Stops the replication task}
\usage{
databasemigrationservice_stop_replication_task(ReplicationTaskArn)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name(ARN) of the replication task to be stopped.}
}
\description{
Stops the replication task.
}
\section{Request syntax}{
\preformatted{svc$stop_replication_task(
  ReplicationTaskArn = "string"
)
}
}

\section{Examples}{
\preformatted{# Stops the replication task.
svc$stop_replication_task(
  ReplicationTaskArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ..."
)
}
}

\keyword{internal}
