% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_cancel_job}
\alias{dataexchange_cancel_job}
\title{This operation cancels a job}
\usage{
dataexchange_cancel_job(JobId)
}
\arguments{
\item{JobId}{[required] The unique identifier for a job.}
}
\description{
This operation cancels a job. Jobs can be cancelled only when they are
in the WAITING state.
}
\section{Request syntax}{
\preformatted{svc$cancel_job(
  JobId = "string"
)
}
}

\keyword{internal}
