% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_create_data_set}
\alias{dataexchange_create_data_set}
\title{This operation creates a data set}
\usage{
dataexchange_create_data_set(AssetType, Description, Name, Tags)
}
\arguments{
\item{AssetType}{[required] The type of file your data is stored in. Currently, the supported asset
type is S3\\_SNAPSHOT.}

\item{Description}{[required] A description for the data set. This value can be up to 16,348
characters long.}

\item{Name}{[required] The name of the data set.}

\item{Tags}{A data set tag is an optional label that you can assign to a data set
when you create it. Each tag consists of a key and an optional value,
both of which you define. When you use tagging, you can also use
tag-based access control in IAM policies to control access to these data
sets and revisions.}
}
\description{
This operation creates a data set.
}
\section{Request syntax}{
\preformatted{svc$create_data_set(
  AssetType = "S3_SNAPSHOT",
  Description = "string",
  Name = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
