% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_create_job}
\alias{dataexchange_create_job}
\title{This operation creates a job}
\usage{
dataexchange_create_job(Details, Type)
}
\arguments{
\item{Details}{[required] The details for the CreateJob request.}

\item{Type}{[required] The type of job to be created.}
}
\description{
This operation creates a job.
}
\section{Request syntax}{
\preformatted{svc$create_job(
  Details = list(
    ExportAssetToSignedUrl = list(
      AssetId = "string",
      DataSetId = "string",
      RevisionId = "string"
    ),
    ExportAssetsToS3 = list(
      AssetDestinations = list(
        list(
          AssetId = "string",
          Bucket = "string",
          Key = "string"
        )
      ),
      DataSetId = "string",
      Encryption = list(
        KmsKeyArn = "string",
        Type = "aws:kms"|"AES256"
      ),
      RevisionId = "string"
    ),
    ImportAssetFromSignedUrl = list(
      AssetName = "string",
      DataSetId = "string",
      Md5Hash = "string",
      RevisionId = "string"
    ),
    ImportAssetsFromS3 = list(
      AssetSources = list(
        list(
          Bucket = "string",
          Key = "string"
        )
      ),
      DataSetId = "string",
      RevisionId = "string"
    )
  ),
  Type = "IMPORT_ASSETS_FROM_S3"|"IMPORT_ASSET_FROM_SIGNED_URL"|"EXPORT_ASSETS_TO_S3"|"EXPORT_ASSET_TO_SIGNED_URL"
)
}
}

\keyword{internal}
