% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_create_revision}
\alias{dataexchange_create_revision}
\title{This operation creates a revision for a data set}
\usage{
dataexchange_create_revision(Comment, DataSetId, Tags)
}
\arguments{
\item{Comment}{An optional comment about the revision.}

\item{DataSetId}{[required] The unique identifier for a data set.}

\item{Tags}{A revision tag is an optional label that you can assign to a revision
when you create it. Each tag consists of a key and an optional value,
both of which you define. When you use tagging, you can also use
tag-based access control in IAM policies to control access to these data
sets and revisions.}
}
\description{
This operation creates a revision for a data set.
}
\section{Request syntax}{
\preformatted{svc$create_revision(
  Comment = "string",
  DataSetId = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
