% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_delete_data_set}
\alias{dataexchange_delete_data_set}
\title{This operation deletes a data set}
\usage{
dataexchange_delete_data_set(DataSetId)
}
\arguments{
\item{DataSetId}{[required] The unique identifier for a data set.}
}
\description{
This operation deletes a data set.
}
\section{Request syntax}{
\preformatted{svc$delete_data_set(
  DataSetId = "string"
)
}
}

\keyword{internal}
