% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_delete_revision}
\alias{dataexchange_delete_revision}
\title{This operation deletes a revision}
\usage{
dataexchange_delete_revision(DataSetId, RevisionId)
}
\arguments{
\item{DataSetId}{[required] The unique identifier for a data set.}

\item{RevisionId}{[required] The unique identifier for a revision.}
}
\description{
This operation deletes a revision.
}
\section{Request syntax}{
\preformatted{svc$delete_revision(
  DataSetId = "string",
  RevisionId = "string"
)
}
}

\keyword{internal}
