% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_get_data_set}
\alias{dataexchange_get_data_set}
\title{This operation returns information about a data set}
\usage{
dataexchange_get_data_set(DataSetId)
}
\arguments{
\item{DataSetId}{[required] The unique identifier for a data set.}
}
\description{
This operation returns information about a data set.
}
\section{Request syntax}{
\preformatted{svc$get_data_set(
  DataSetId = "string"
)
}
}

\keyword{internal}
