% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_get_job}
\alias{dataexchange_get_job}
\title{This operation returns information about a job}
\usage{
dataexchange_get_job(JobId)
}
\arguments{
\item{JobId}{[required] The unique identifier for a job.}
}
\description{
This operation returns information about a job.
}
\section{Request syntax}{
\preformatted{svc$get_job(
  JobId = "string"
)
}
}

\keyword{internal}
