% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_list_data_set_revisions}
\alias{dataexchange_list_data_set_revisions}
\title{This operation lists a data set's revisions sorted by CreatedAt in
descending order}
\usage{
dataexchange_list_data_set_revisions(DataSetId, MaxResults, NextToken)
}
\arguments{
\item{DataSetId}{[required] The unique identifier for a data set.}

\item{MaxResults}{The maximum number of results returned by a single call.}

\item{NextToken}{The token value retrieved from a previous call to access the next page
of results.}
}
\description{
This operation lists a data set's revisions sorted by CreatedAt in
descending order.
}
\section{Request syntax}{
\preformatted{svc$list_data_set_revisions(
  DataSetId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
