% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_list_data_sets}
\alias{dataexchange_list_data_sets}
\title{This operation lists your data sets}
\usage{
dataexchange_list_data_sets(MaxResults, NextToken, Origin)
}
\arguments{
\item{MaxResults}{The maximum number of results returned by a single call.}

\item{NextToken}{The token value retrieved from a previous call to access the next page
of results.}

\item{Origin}{A property that defines the data set as OWNED by the account (for
providers) or ENTITLED to the account (for subscribers).}
}
\description{
This operation lists your data sets. When listing by origin OWNED,
results are sorted by CreatedAt in descending order. When listing by
origin ENTITLED, there is no order and the maxResults parameter is
ignored.
}
\section{Request syntax}{
\preformatted{svc$list_data_sets(
  MaxResults = 123,
  NextToken = "string",
  Origin = "string"
)
}
}

\keyword{internal}
