% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_list_jobs}
\alias{dataexchange_list_jobs}
\title{This operation lists your jobs sorted by CreatedAt in descending order}
\usage{
dataexchange_list_jobs(DataSetId, MaxResults, NextToken, RevisionId)
}
\arguments{
\item{DataSetId}{The unique identifier for a data set.}

\item{MaxResults}{The maximum number of results returned by a single call.}

\item{NextToken}{The token value retrieved from a previous call to access the next page
of results.}

\item{RevisionId}{The unique identifier for a revision.}
}
\description{
This operation lists your jobs sorted by CreatedAt in descending order.
}
\section{Request syntax}{
\preformatted{svc$list_jobs(
  DataSetId = "string",
  MaxResults = 123,
  NextToken = "string",
  RevisionId = "string"
)
}
}

\keyword{internal}
