% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_list_revision_assets}
\alias{dataexchange_list_revision_assets}
\title{This operation lists a revision's assets sorted alphabetically in
descending order}
\usage{
dataexchange_list_revision_assets(DataSetId, MaxResults, NextToken,
  RevisionId)
}
\arguments{
\item{DataSetId}{[required] The unique identifier for a data set.}

\item{MaxResults}{The maximum number of results returned by a single call.}

\item{NextToken}{The token value retrieved from a previous call to access the next page
of results.}

\item{RevisionId}{[required] The unique identifier for a revision.}
}
\description{
This operation lists a revision\'s assets sorted alphabetically in
descending order.
}
\section{Request syntax}{
\preformatted{svc$list_revision_assets(
  DataSetId = "string",
  MaxResults = 123,
  NextToken = "string",
  RevisionId = "string"
)
}
}

\keyword{internal}
