% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_list_tags_for_resource}
\alias{dataexchange_list_tags_for_resource}
\title{This operation lists the tags on the resource}
\usage{
dataexchange_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies an AWS resource.}
}
\description{
This operation lists the tags on the resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
