% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_start_job}
\alias{dataexchange_start_job}
\title{This operation starts a job}
\usage{
dataexchange_start_job(JobId)
}
\arguments{
\item{JobId}{[required] The unique identifier for a job.}
}
\description{
This operation starts a job.
}
\section{Request syntax}{
\preformatted{svc$start_job(
  JobId = "string"
)
}
}

\keyword{internal}
