% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_tag_resource}
\alias{dataexchange_tag_resource}
\title{This operation tags a resource}
\usage{
dataexchange_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies an AWS resource.}

\item{Tags}{[required] A label that consists of a customer-defined key and an optional value.}
}
\description{
This operation tags a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
