% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_untag_resource}
\alias{dataexchange_untag_resource}
\title{This operation removes one or more tags from a resource}
\usage{
dataexchange_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies an AWS resource.}

\item{TagKeys}{[required] The key tags.}
}
\description{
This operation removes one or more tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
