% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_update_asset}
\alias{dataexchange_update_asset}
\title{This operation updates an asset}
\usage{
dataexchange_update_asset(AssetId, DataSetId, Name, RevisionId)
}
\arguments{
\item{AssetId}{[required] The unique identifier for an asset.}

\item{DataSetId}{[required] The unique identifier for a data set.}

\item{Name}{[required] The name of the asset. When importing from Amazon S3, the S3 object key
is used as the asset name. When exporting to Amazon S3, the asset name
is used as default target S3 object key.}

\item{RevisionId}{[required] The unique identifier for a revision.}
}
\description{
This operation updates an asset.
}
\section{Request syntax}{
\preformatted{svc$update_asset(
  AssetId = "string",
  DataSetId = "string",
  Name = "string",
  RevisionId = "string"
)
}
}

\keyword{internal}
