% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_update_data_set}
\alias{dataexchange_update_data_set}
\title{This operation updates a data set}
\usage{
dataexchange_update_data_set(DataSetId, Description, Name)
}
\arguments{
\item{DataSetId}{[required] The unique identifier for a data set.}

\item{Description}{The description for the data set.}

\item{Name}{The name of the data set.}
}
\description{
This operation updates a data set.
}
\section{Request syntax}{
\preformatted{svc$update_data_set(
  DataSetId = "string",
  Description = "string",
  Name = "string"
)
}
}

\keyword{internal}
