% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexchange_operations.R
\name{dataexchange_update_revision}
\alias{dataexchange_update_revision}
\title{This operation updates a revision}
\usage{
dataexchange_update_revision(Comment, DataSetId, Finalized, RevisionId)
}
\arguments{
\item{Comment}{An optional comment about the revision.}

\item{DataSetId}{[required] The unique identifier for a data set.}

\item{Finalized}{Finalizing a revision tells AWS Data Exchange that your changes to the
assets in the revision are complete. After it's in this read-only state,
you can publish the revision to your products.}

\item{RevisionId}{[required] The unique identifier for a revision.}
}
\description{
This operation updates a revision.
}
\section{Request syntax}{
\preformatted{svc$update_revision(
  Comment = "string",
  DataSetId = "string",
  Finalized = TRUE|FALSE,
  RevisionId = "string"
)
}
}

\keyword{internal}
