% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapipeline_operations.R
\name{datapipeline_list_pipelines}
\alias{datapipeline_list_pipelines}
\title{Lists the pipeline identifiers for all active pipelines that you have
permission to access}
\usage{
datapipeline_list_pipelines(marker)
}
\arguments{
\item{marker}{The starting point for the results to be returned. For the first call,
this value should be empty. As long as there are more results, continue
to call \code{ListPipelines} with the marker value from the previous call to
retrieve the next set of results.}
}
\description{
Lists the pipeline identifiers for all active pipelines that you have
permission to access.
}
\section{Request syntax}{
\preformatted{svc$list_pipelines(
  marker = "string"
)
}
}

\keyword{internal}
