% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_location_smb}
\alias{datasync_create_location_smb}
\title{Defines a file system on an Server Message Block (SMB) server that can
be read from or written to}
\usage{
datasync_create_location_smb(Subdirectory, ServerHostname, User, Domain,
  Password, AgentArns, MountOptions, Tags)
}
\arguments{
\item{Subdirectory}{[required] The subdirectory in the SMB file system that is used to read data from
the SMB source location or write data to the SMB destination. The SMB
path should be a path that\'s exported by the SMB server, or a
subdirectory of that path. The path should be such that it can be
mounted by other SMB clients in your network.

\code{Subdirectory} must be specified with forward slashes. For example
\verb{/path/to/folder}.

To transfer all the data in the folder you specified, DataSync needs to
have permissions to mount the SMB share, as well as to access all the
data in that share. To ensure this, either ensure that the user/password
specified belongs to the user who can mount the share, and who has the
appropriate permissions for all of the files and directories that you
want DataSync to access, or use credentials of a member of the Backup
Operators group to mount the share. Doing either enables the agent to
access the data. For the agent to access directories, you must
additionally enable all execute access.}

\item{ServerHostname}{[required] The name of the SMB server. This value is the IP address or Domain Name
Service (DNS) name of the SMB server. An agent that is installed
on-premises uses this hostname to mount the SMB server in a network.

This name must either be DNS-compliant or must be an IP version 4 (IPv4)
address.}

\item{User}{[required] The user who can mount the share, has the permissions to access files
and folders in the SMB share.}

\item{Domain}{The name of the Windows domain that the SMB server belongs to.}

\item{Password}{[required] The password of the user who can mount the share, has the permissions to
access files and folders in the SMB share.}

\item{AgentArns}{[required] The Amazon Resource Names (ARNs) of agents to use for a Simple Message
Block (SMB) location.}

\item{MountOptions}{The mount options used by DataSync to access the SMB server.}

\item{Tags}{The key-value pair that represents the tag that you want to add to the
location. The value can be an empty string. We recommend using tags to
name your resources.}
}
\description{
Defines a file system on an Server Message Block (SMB) server that can
be read from or written to.
}
\section{Request syntax}{
\preformatted{svc$create_location_smb(
  Subdirectory = "string",
  ServerHostname = "string",
  User = "string",
  Domain = "string",
  Password = "string",
  AgentArns = list(
    "string"
  ),
  MountOptions = list(
    Version = "AUTOMATIC"|"SMB2"|"SMB3"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
