% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_task}
\alias{datasync_create_task}
\title{Creates a task}
\usage{
datasync_create_task(SourceLocationArn, DestinationLocationArn,
  CloudWatchLogGroupArn, Name, Options, Excludes, Schedule, Tags)
}
\arguments{
\item{SourceLocationArn}{[required] The Amazon Resource Name (ARN) of the source location for the task.}

\item{DestinationLocationArn}{[required] The Amazon Resource Name (ARN) of an AWS storage resource's location.}

\item{CloudWatchLogGroupArn}{The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that
is used to monitor and log events in the task.}

\item{Name}{The name of a task. This value is a text reference that is used to
identify the task in the console.}

\item{Options}{The set of configuration options that control the behavior of a single
execution of the task that occurs when you call
\code{\link[=datasync_start_task_execution]{start_task_execution}}. You can
configure these options to preserve metadata such as user ID (UID) and
group ID (GID), file permissions, data integrity verification, and so
on.

For each individual task execution, you can override these options by
specifying the \code{OverrideOptions} before starting the task execution. For
more information, see the operation.}

\item{Excludes}{A list of filter rules that determines which files to exclude from a
task. The list should contain a single filter string that consists of
the patterns to exclude. The patterns are delimited by "|" (that is, a
pipe), for example, \code{"/folder1|/folder2"}.}

\item{Schedule}{Specifies a schedule used to periodically transfer files from a source
to a destination location. The schedule should be specified in UTC time.
For more information, see task-scheduling.}

\item{Tags}{The key-value pair that represents the tag that you want to add to the
resource. The value can be an empty string.}
}
\description{
Creates a task. A task is a set of two locations (source and
destination) and a set of Options that you use to control the behavior
of a task. If you don't specify Options when you create a task, AWS
DataSync populates them with service defaults.

When you create a task, it first enters the CREATING state. During
CREATING AWS DataSync attempts to mount the on-premises Network File
System (NFS) location. The task transitions to the AVAILABLE state
without waiting for the AWS location to become mounted. If required, AWS
DataSync mounts the AWS location before each task execution.

If an agent that is associated with a source (NFS) location goes
offline, the task transitions to the UNAVAILABLE status. If the status
of the task remains in the CREATING status for more than a few minutes,
it means that your agent might be having trouble mounting the source NFS
file system. Check the task's ErrorCode and ErrorDetail. Mount issues
are often caused by either a misconfigured firewall or a mistyped NFS
server hostname.
}
\section{Request syntax}{
\preformatted{svc$create_task(
  SourceLocationArn = "string",
  DestinationLocationArn = "string",
  CloudWatchLogGroupArn = "string",
  Name = "string",
  Options = list(
    VerifyMode = "POINT_IN_TIME_CONSISTENT"|"ONLY_FILES_TRANSFERRED"|"NONE",
    OverwriteMode = "ALWAYS"|"NEVER",
    Atime = "NONE"|"BEST_EFFORT",
    Mtime = "NONE"|"PRESERVE",
    Uid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    Gid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    PreserveDeletedFiles = "PRESERVE"|"REMOVE",
    PreserveDevices = "NONE"|"PRESERVE",
    PosixPermissions = "NONE"|"PRESERVE",
    BytesPerSecond = 123,
    TaskQueueing = "ENABLED"|"DISABLED",
    LogLevel = "OFF"|"BASIC"|"TRANSFER",
    TransferMode = "CHANGED"|"ALL"
  ),
  Excludes = list(
    list(
      FilterType = "SIMPLE_PATTERN",
      Value = "string"
    )
  ),
  Schedule = list(
    ScheduleExpression = "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
