% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_location_smb}
\alias{datasync_describe_location_smb}
\title{Returns metadata, such as the path and user information about a SMB
location}
\usage{
datasync_describe_location_smb(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon resource Name (ARN) of the SMB location to describe.}
}
\description{
Returns metadata, such as the path and user information about a SMB
location.
}
\section{Request syntax}{
\preformatted{svc$describe_location_smb(
  LocationArn = "string"
)
}
}

\keyword{internal}
