% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_list_locations}
\alias{datasync_list_locations}
\title{Returns a lists of source and destination locations}
\usage{
datasync_list_locations(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of locations to return.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next
list of locations.}
}
\description{
Returns a lists of source and destination locations.
}
\details{
If you have more locations than are returned in a response (that is, the
response returns only a truncated list of your agents), the response
contains a token that you can specify in your next request to fetch the
next page of locations.
}
\section{Request syntax}{
\preformatted{datasync$list_locations(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
