% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_list_tasks}
\alias{datasync_list_tasks}
\title{Returns a list of all the tasks}
\usage{
datasync_list_tasks(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of tasks to return.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next
list of tasks.}
}
\description{
Returns a list of all the tasks.
}
\section{Request syntax}{
\preformatted{svc$list_tasks(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
