% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_start_task_execution}
\alias{datasync_start_task_execution}
\title{Starts a specific invocation of a task}
\usage{
datasync_start_task_execution(TaskArn, OverrideOptions, Includes)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the task to start.}

\item{OverrideOptions}{}

\item{Includes}{A list of filter rules that determines which files to include when
running a task. The pattern should contain a single filter string that
consists of the patterns to include. The patterns are delimited by "|"
(that is, a pipe). For example: \code{"/folder1|/folder2"}}
}
\description{
Starts a specific invocation of a task. A \code{TaskExecution} value
represents an individual run of a task. Each task can have at most one
\code{TaskExecution} at a time.

\code{TaskExecution} has the following transition phases: INITIALIZING |
PREPARING | TRANSFERRING | VERIFYING | SUCCESS/FAILURE.

For detailed information, see the Task Execution section in the
Components and Terminology topic in the \emph{AWS DataSync User Guide}.
}
\section{Request syntax}{
\preformatted{svc$start_task_execution(
  TaskArn = "string",
  OverrideOptions = list(
    VerifyMode = "POINT_IN_TIME_CONSISTENT"|"ONLY_FILES_TRANSFERRED"|"NONE",
    OverwriteMode = "ALWAYS"|"NEVER",
    Atime = "NONE"|"BEST_EFFORT",
    Mtime = "NONE"|"PRESERVE",
    Uid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    Gid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    PreserveDeletedFiles = "PRESERVE"|"REMOVE",
    PreserveDevices = "NONE"|"PRESERVE",
    PosixPermissions = "NONE"|"PRESERVE",
    BytesPerSecond = 123,
    TaskQueueing = "ENABLED"|"DISABLED",
    LogLevel = "OFF"|"BASIC"|"TRANSFER",
    TransferMode = "CHANGED"|"ALL"
  ),
  Includes = list(
    list(
      FilterType = "SIMPLE_PATTERN",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
