% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_tag_resource}
\alias{datasync_tag_resource}
\title{Applies a key-value pair to an AWS resource}
\usage{
datasync_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to apply the tag to.}

\item{Tags}{[required] The tags to apply.}
}
\description{
Applies a key-value pair to an AWS resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
