% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_update_task}
\alias{datasync_update_task}
\title{Updates the metadata associated with a task}
\usage{
datasync_update_task(TaskArn, Options, Excludes, Name,
  CloudWatchLogGroupArn)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the resource name of the task to
update.}

\item{Options}{}

\item{Excludes}{A filter that determines which files to exclude from a task based on the
specified pattern in the filter. Transfers all files in the task's
subdirectory, except files that match the filter that is set.}

\item{Name}{The name of the task to update.}

\item{CloudWatchLogGroupArn}{The Amazon Resource Name (ARN) of the resource name of the CloudWatch
LogGroup.}
}
\description{
Updates the metadata associated with a task.
}
\section{Request syntax}{
\preformatted{svc$update_task(
  TaskArn = "string",
  Options = list(
    VerifyMode = "POINT_IN_TIME_CONSISTENT"|"NONE",
    Atime = "NONE"|"BEST_EFFORT",
    Mtime = "NONE"|"PRESERVE",
    Uid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    Gid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    PreserveDeletedFiles = "PRESERVE"|"REMOVE",
    PreserveDevices = "NONE"|"PRESERVE",
    PosixPermissions = "NONE"|"BEST_EFFORT"|"PRESERVE",
    BytesPerSecond = 123
  ),
  Excludes = list(
    list(
      FilterType = "SIMPLE_PATTERN",
      Value = "string"
    )
  ),
  Name = "string",
  CloudWatchLogGroupArn = "string"
)
}
}

\keyword{internal}
