% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_update_task_execution}
\alias{datasync_update_task_execution}
\title{Updates execution of a task}
\usage{
datasync_update_task_execution(TaskExecutionArn, Options)
}
\arguments{
\item{TaskExecutionArn}{[required] The Amazon Resource Name (ARN) of the specific task execution that is
being updated.}

\item{Options}{[required]}
}
\description{
Updates execution of a task.

You can modify bandwidth throttling for a task execution that is running
or queued. For more information, see Adjusting Bandwidth Throttling for
a Task Execution.

The only \code{Option} that can be modified by
\code{\link[=datasync_update_task_execution]{update_task_execution}} is
\code{BytesPerSecond}.
}
\section{Request syntax}{
\preformatted{svc$update_task_execution(
  TaskExecutionArn = "string",
  Options = list(
    VerifyMode = "POINT_IN_TIME_CONSISTENT"|"ONLY_FILES_TRANSFERRED"|"NONE",
    OverwriteMode = "ALWAYS"|"NEVER",
    Atime = "NONE"|"BEST_EFFORT",
    Mtime = "NONE"|"PRESERVE",
    Uid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    Gid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    PreserveDeletedFiles = "PRESERVE"|"REMOVE",
    PreserveDevices = "NONE"|"PRESERVE",
    PosixPermissions = "NONE"|"PRESERVE",
    BytesPerSecond = 123,
    TaskQueueing = "ENABLED"|"DISABLED",
    LogLevel = "OFF"|"BASIC"|"TRANSFER",
    TransferMode = "CHANGED"|"ALL"
  )
)
}
}

\keyword{internal}
