% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{default_parameters}
\alias{default_parameters}
\title{Default parameters for extracting paw features}
\usage{
default_parameters()
}
\value{
parameters used for Jones et al. (2020)
}
\description{
Default parameters for extracting paw features
}
