% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{default_standards}
\alias{default_standards}
\title{Default strain-based standards for paw features}
\usage{
default_standards()
}
\value{
standards used for Jones et al. (2020)
}
\description{
Default strain-based standards for paw features
}
