% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_create_graph}
\alias{detective_create_graph}
\title{Amazon Detective is currently in preview}
\usage{
detective_create_graph()
}
\description{
Amazon Detective is currently in preview.
}
\details{
Creates a new behavior graph for the calling account, and sets that
account as the master account. This operation is called by the account
that is enabling Detective.

The operation also enables Detective for the calling account in the
currently selected Region. It returns the ARN of the new behavior graph.

\code{CreateGraph} triggers a process to create the corresponding data tables
for the new behavior graph.

An account can only be the master account for one behavior graph within
a Region. If the same account calls \code{CreateGraph} with the same master
account, it always returns the same behavior graph ARN. It does not
create a new behavior graph.
}
\keyword{internal}
