% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_create_members}
\alias{detective_create_members}
\title{Sends a request to invite the specified AWS accounts to be member
accounts in the behavior graph}
\usage{
detective_create_members(GraphArn, Message, Accounts)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph to invite the member accounts to
contribute their data to.}

\item{Message}{Customized message text to include in the invitation email message to
the invited member accounts.}

\item{Accounts}{[required] The list of AWS accounts to invite to become member accounts in the
behavior graph. For each invited account, the account list contains the
account identifier and the AWS account root user email address.}
}
\description{
Sends a request to invite the specified AWS accounts to be member
accounts in the behavior graph. This operation can only be called by the
master account for a behavior graph.

\code{CreateMembers} verifies the accounts and then sends invitations to the
verified accounts.

The request provides the behavior graph ARN and the list of accounts to
invite.

The response separates the requested accounts into two lists:
\itemize{
\item The accounts that \code{CreateMembers} was able to start the verification
for. This list includes member accounts that are being verified,
that have passed verification and are being sent an invitation, and
that have failed verification.
\item The accounts that \code{CreateMembers} was unable to process. This list
includes accounts that were already invited to be member accounts in
the behavior graph.
}
}
\section{Request syntax}{
\preformatted{svc$create_members(
  GraphArn = "string",
  Message = "string",
  Accounts = list(
    list(
      AccountId = "string",
      EmailAddress = "string"
    )
  )
)
}
}

\keyword{internal}
