% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_delete_members}
\alias{detective_delete_members}
\title{Amazon Detective is currently in preview}
\usage{
detective_delete_members(GraphArn, AccountIds)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph to delete members from.}

\item{AccountIds}{[required] The list of AWS account identifiers for the member accounts to delete
from the behavior graph.}
}
\description{
Amazon Detective is currently in preview.
}
\details{
Deletes one or more member accounts from the master account behavior
graph. This operation can only be called by a Detective master account.
That account cannot use \code{DeleteMembers} to delete their own account from
the behavior graph. To disable a behavior graph, the master account uses
the \code{DeleteGraph} API method.
}
\section{Request syntax}{
\preformatted{svc$delete_members(
  GraphArn = "string",
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
