% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_list_graphs}
\alias{detective_list_graphs}
\title{Returns the list of behavior graphs that the calling account is a master
of}
\usage{
detective_list_graphs(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{For requests to get the next page of results, the pagination token that
was returned with the previous set of results. The initial request does
not include a pagination token.}

\item{MaxResults}{The maximum number of graphs to return at a time. The total must be less
than the overall limit on the number of results to return, which is
currently 200.}
}
\description{
Returns the list of behavior graphs that the calling account is a master
of. This operation can only be called by a master account.

Because an account can currently only be the master of one behavior
graph within a Region, the results always contain a single graph.
}
\section{Request syntax}{
\preformatted{svc$list_graphs(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
