% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_list_invitations}
\alias{detective_list_invitations}
\title{Retrieves the list of open and accepted behavior graph invitations for
the member account}
\usage{
detective_list_invitations(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{For requests to retrieve the next page of results, the pagination token
that was returned with the previous page of results. The initial request
does not include a pagination token.}

\item{MaxResults}{The maximum number of behavior graph invitations to return in the
response. The total must be less than the overall limit on the number of
results to return, which is currently 200.}
}
\description{
Retrieves the list of open and accepted behavior graph invitations for
the member account. This operation can only be called by a member
account.

Open invitations are invitations that the member account has not
responded to.

The results do not include behavior graphs for which the member account
declined the invitation. The results also do not include behavior graphs
that the member account resigned from or was removed from.
}
\section{Request syntax}{
\preformatted{svc$list_invitations(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
