% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_list_members}
\alias{detective_list_members}
\title{Amazon Detective is currently in preview}
\usage{
detective_list_members(GraphArn, NextToken, MaxResults)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph for which to retrieve the list of member
accounts.}

\item{NextToken}{For requests to retrieve the next page of member account results, the
pagination token that was returned with the previous page of results.
The initial request does not include a pagination token.}

\item{MaxResults}{The maximum number of member accounts to include in the response. The
total must be less than the overall limit on the number of results to
return, which is currently 200.}
}
\description{
Amazon Detective is currently in preview.
}
\details{
Retrieves the list of member accounts for a behavior graph. Does not
return member accounts that were removed from the behavior graph.
}
\section{Request syntax}{
\preformatted{svc$list_members(
  GraphArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
