% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_reject_invitation}
\alias{detective_reject_invitation}
\title{Amazon Detective is currently in preview}
\usage{
detective_reject_invitation(GraphArn)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph to reject the invitation to.

The member account\'s current member status in the behavior graph must
be \code{INVITED}.}
}
\description{
Amazon Detective is currently in preview.
}
\details{
Rejects an invitation to contribute the account data to a behavior
graph. This operation must be called by a member account that has the
\code{INVITED} status.
}
\section{Request syntax}{
\preformatted{svc$reject_invitation(
  GraphArn = "string"
)
}
}

\keyword{internal}
