% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_start_monitoring_member}
\alias{detective_start_monitoring_member}
\title{Sends a request to enable data ingest for a member account that has a
status of ACCEPTED_BUT_DISABLED}
\usage{
detective_start_monitoring_member(GraphArn, AccountId)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph.}

\item{AccountId}{[required] The account ID of the member account to try to enable.

The account must be an invited member account with a status of
\code{ACCEPTED_BUT_DISABLED}.}
}
\description{
Sends a request to enable data ingest for a member account that has a
status of \code{ACCEPTED_BUT_DISABLED}.

For valid member accounts, the status is updated as follows.
\itemize{
\item If Detective enabled the member account, then the new status is
\code{ENABLED}.
\item If Detective cannot enable the member account, the status remains
\code{ACCEPTED_BUT_DISABLED}.
}
}
\section{Request syntax}{
\preformatted{svc$start_monitoring_member(
  GraphArn = "string",
  AccountId = "string"
)
}
}

\keyword{internal}
