% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_create_device_pool}
\alias{devicefarm_create_device_pool}
\title{Creates a device pool}
\usage{
devicefarm_create_device_pool(projectArn, name, description, rules,
  maxDevices)
}
\arguments{
\item{projectArn}{[required] The ARN of the project for the device pool.}

\item{name}{[required] The device pool\'s name.}

\item{description}{The device pool\'s description.}

\item{rules}{[required] The device pool\'s rules.}

\item{maxDevices}{The number of devices that Device Farm can add to your device pool.
Device Farm adds devices that are available and meet the criteria that
you assign for the \code{rules} parameter. Depending on how many devices meet
these constraints, your device pool might contain fewer devices than the
value for this parameter.

By specifying the maximum number of devices, you can control the costs
that you incur by running tests.}
}
\description{
Creates a device pool.
}
\section{Request syntax}{
\preformatted{svc$create_device_pool(
  projectArn = "string",
  name = "string",
  description = "string",
  rules = list(
    list(
      attribute = "ARN"|"PLATFORM"|"FORM_FACTOR"|"MANUFACTURER"|"REMOTE_ACCESS_ENABLED"|"REMOTE_DEBUG_ENABLED"|"APPIUM_VERSION"|"INSTANCE_ARN"|"INSTANCE_LABELS"|"FLEET_TYPE"|"OS_VERSION"|"MODEL"|"AVAILABILITY",
      operator = "EQUALS"|"LESS_THAN"|"LESS_THAN_OR_EQUALS"|"GREATER_THAN"|"GREATER_THAN_OR_EQUALS"|"IN"|"NOT_IN"|"CONTAINS",
      value = "string"
    )
  ),
  maxDevices = 123
)
}
}

\section{Examples}{
\preformatted{# The following example creates a new device pool named MyDevicePool
# inside an existing project.
svc$create_device_pool(
  name = "MyDevicePool",
  description = "My Android devices",
  projectArn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  rules = list()
)
}
}

\keyword{internal}
