% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_create_test_grid_url}
\alias{devicefarm_create_test_grid_url}
\title{Creates a signed, short-term URL that can be passed to a Selenium
RemoteWebDriver constructor}
\usage{
devicefarm_create_test_grid_url(projectArn, expiresInSeconds)
}
\arguments{
\item{projectArn}{[required] ARN (from CreateTestGridProject or ListTestGridProjects) to associate
with the short-term URL.}

\item{expiresInSeconds}{[required] Lifetime, in seconds, of the URL.}
}
\description{
Creates a signed, short-term URL that can be passed to a Selenium
\code{RemoteWebDriver} constructor.
}
\section{Request syntax}{
\preformatted{svc$create_test_grid_url(
  projectArn = "string",
  expiresInSeconds = 123
)
}
}

\keyword{internal}
