% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_upload}
\alias{devicefarm_delete_upload}
\title{Deletes an upload given the upload ARN}
\usage{
devicefarm_delete_upload(arn)
}
\arguments{
\item{arn}{[required] Represents the Amazon Resource Name (ARN) of the Device Farm upload you
wish to delete.}
}
\description{
Deletes an upload given the upload ARN.
}
\section{Request syntax}{
\preformatted{devicefarm$delete_upload(
  arn = "string"
)
}
}

\examples{
# The following example deletes a specific upload.
\donttest{devicefarm$delete_upload(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:upload:EXAMPLE-GUID-123-456"
)}

}
\keyword{internal}
