% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_device_instance}
\alias{devicefarm_get_device_instance}
\title{Returns information about a device instance that belongs to a private
device fleet}
\usage{
devicefarm_get_device_instance(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the instance you\'re requesting
information about.}
}
\description{
Returns information about a device instance that belongs to a private
device fleet.
}
\section{Request syntax}{
\preformatted{svc$get_device_instance(
  arn = "string"
)
}
}

\keyword{internal}
