% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_instance_profile}
\alias{devicefarm_get_instance_profile}
\title{Returns information about the specified instance profile}
\usage{
devicefarm_get_instance_profile(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of your instance profile.}
}
\description{
Returns information about the specified instance profile.
}
\section{Request syntax}{
\preformatted{devicefarm$get_instance_profile(
  arn = "string"
)
}
}

\keyword{internal}
