% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_project}
\alias{devicefarm_get_project}
\title{Gets information about a project}
\usage{
devicefarm_get_project(arn)
}
\arguments{
\item{arn}{[required] The project's ARN.}
}
\description{
Gets information about a project.
}
\section{Request syntax}{
\preformatted{devicefarm$get_project(
  arn = "string"
)
}
}

\examples{
# The following example gets information about a specific project.
\donttest{devicefarm$get_project(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:5e01a8c7-c861-4c0a-b1d5-12345EXAM..."
)}

}
\keyword{internal}
