% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_remote_access_session}
\alias{devicefarm_get_remote_access_session}
\title{Returns a link to a currently running remote access session}
\usage{
devicefarm_get_remote_access_session(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the remote access session about which
you want to get session information.}
}
\description{
Returns a link to a currently running remote access session.
}
\section{Request syntax}{
\preformatted{svc$get_remote_access_session(
  arn = "string"
)
}
}

\section{Examples}{
\preformatted{# The following example gets a specific remote access session.
svc$get_remote_access_session(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:session:EXAMPLE-GUID-123-456"
)
}
}

\keyword{internal}
