% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_suite}
\alias{devicefarm_get_suite}
\title{Gets information about a suite}
\usage{
devicefarm_get_suite(arn)
}
\arguments{
\item{arn}{[required] The suite's ARN.}
}
\description{
Gets information about a suite.
}
\section{Request syntax}{
\preformatted{svc$get_suite(
  arn = "string"
)
}
}

\examples{
\dontrun{
# The following example gets information about a specific test suite.
svc$get_suite(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:suite:EXAMPLE-GUID-123-456"
)
}

}
\keyword{internal}
