% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_test}
\alias{devicefarm_get_test}
\title{Gets information about a test}
\usage{
devicefarm_get_test(arn)
}
\arguments{
\item{arn}{[required] The test\'s ARN.}
}
\description{
Gets information about a test.
}
\section{Request syntax}{
\preformatted{svc$get_test(
  arn = "string"
)
}
}

\examples{
# The following example gets information about a specific test.
\dontrun{svc$get_test(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:test:EXAMPLE-GUID-123-456"
)}

}
\keyword{internal}
