% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_test_grid_project}
\alias{devicefarm_get_test_grid_project}
\title{Retrieves information about a Selenium testing project}
\usage{
devicefarm_get_test_grid_project(projectArn)
}
\arguments{
\item{projectArn}{[required] The ARN of the Selenium testing project, from either
CreateTestGridProject or ListTestGridProjects.}
}
\description{
Retrieves information about a Selenium testing project.
}
\section{Request syntax}{
\preformatted{svc$get_test_grid_project(
  projectArn = "string"
)
}
}

\keyword{internal}
