% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_test_grid_session}
\alias{devicefarm_get_test_grid_session}
\title{A session is an instance of a browser created through a RemoteWebDriver
with the URL from CreateTestGridUrlResult$url}
\usage{
devicefarm_get_test_grid_session(projectArn, sessionId, sessionArn)
}
\arguments{
\item{projectArn}{The ARN for the project that this session belongs to. See
CreateTestGridProject and ListTestGridProjects.}

\item{sessionId}{An ID associated with this session.}

\item{sessionArn}{An ARN that uniquely identifies a TestGridSession.}
}
\description{
A session is an instance of a browser created through a
\code{RemoteWebDriver} with the URL from CreateTestGridUrlResult$url. You can
use the following to look up sessions:
}
\details{
\itemize{
\item The session ARN (GetTestGridSessionRequest$sessionArn).
\item The project ARN and a session ID
(GetTestGridSessionRequest$projectArn and
GetTestGridSessionRequest$sessionId).
}
}
\section{Request syntax}{
\preformatted{svc$get_test_grid_session(
  projectArn = "string",
  sessionId = "string",
  sessionArn = "string"
)
}
}

\keyword{internal}
