% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_upload}
\alias{devicefarm_get_upload}
\title{Gets information about an upload}
\usage{
devicefarm_get_upload(arn)
}
\arguments{
\item{arn}{[required] The upload's ARN.}
}
\description{
Gets information about an upload.
}
\section{Request syntax}{
\preformatted{svc$get_upload(
  arn = "string"
)
}
}

\examples{
\dontrun{
# The following example gets information about a specific upload.
svc$get_upload(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:upload:EXAMPLE-GUID-123-456"
)
}

}
\keyword{internal}
