% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_artifacts}
\alias{devicefarm_list_artifacts}
\title{Gets information about artifacts}
\usage{
devicefarm_list_artifacts(arn, type, nextToken)
}
\arguments{
\item{arn}{[required] The Run, Job, Suite, or Test ARN.}

\item{type}{[required] The artifacts\' type.

Allowed values include:
\itemize{
\item FILE: The artifacts are files.
\item LOG: The artifacts are logs.
\item SCREENSHOT: The artifacts are screenshots.
}}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Gets information about artifacts.
}
\section{Request syntax}{
\preformatted{svc$list_artifacts(
  arn = "string",
  type = "SCREENSHOT"|"FILE"|"LOG",
  nextToken = "string"
)
}
}

\examples{
# The following example lists screenshot artifacts for a specific run.
\donttest{svc$list_artifacts(
  type = "SCREENSHOT",
  arn = "arn:aws:devicefarm:us-west-2:123456789101:run:EXAMPLE-GUID-123-456"
)}

}
\keyword{internal}
