% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_device_pools}
\alias{devicefarm_list_device_pools}
\title{Gets information about device pools}
\usage{
devicefarm_list_device_pools(arn, type, nextToken)
}
\arguments{
\item{arn}{[required] The project ARN.}

\item{type}{The device pools\' type.

Allowed values include:
\itemize{
\item CURATED: A device pool that is created and managed by AWS Device
Farm.
\item PRIVATE: A device pool that is created and managed by the device
pool developer.
}}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Gets information about device pools.
}
\section{Request syntax}{
\preformatted{svc$list_device_pools(
  arn = "string",
  type = "CURATED"|"PRIVATE",
  nextToken = "string"
)
}
}

\examples{
# The following example returns information about the private device pools
# in a specific project.
\donttest{svc$list_device_pools(
  type = "PRIVATE",
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)}

}
\keyword{internal}
