% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_devices}
\alias{devicefarm_list_devices}
\title{Gets information about unique device types}
\usage{
devicefarm_list_devices(arn, nextToken, filters)
}
\arguments{
\item{arn}{The Amazon Resource Name (ARN) of the project.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{filters}{Used to select a set of devices. A filter is made up of an attribute, an
operator, and one or more values.
\itemize{
\item Attribute: The aspect of a device such as platform or model used as
the selection criteria in a device filter.

Allowed values include:
\itemize{
\item ARN: The Amazon Resource Name (ARN) of the device (for example,
\verb{arn:aws:devicefarm:us-west-2::device:12345Example}).
\item PLATFORM: The device platform. Valid values are ANDROID or IOS.
\item OS\\_VERSION: The operating system version (for example, 10.3.2).
\item MODEL: The device model (for example, iPad 5th Gen).
\item AVAILABILITY: The current availability of the device. Valid
values are AVAILABLE, HIGHLY\\_AVAILABLE, BUSY, or
TEMPORARY\\_NOT\\_AVAILABLE.
\item FORM\\_FACTOR: The device form factor. Valid values are PHONE or
TABLET.
\item MANUFACTURER: The device manufacturer (for example, Apple).
\item REMOTE\\_ACCESS\\_ENABLED: Whether the device is enabled for
remote access. Valid values are TRUE or FALSE.
\item REMOTE\\_DEBUG\\_ENABLED: Whether the device is enabled for remote
debugging. Valid values are TRUE or FALSE. Because remote
debugging is \href{https://docs.aws.amazon.com/devicefarm/latest/developerguide/history.html}{no longer supported},
this attribute is ignored.
\item INSTANCE\\_ARN: The Amazon Resource Name (ARN) of the device
instance.
\item INSTANCE\\_LABELS: The label of the device instance.
\item FLEET\\_TYPE: The fleet type. Valid values are PUBLIC or PRIVATE.
}
\item Operator: The filter operator.
\itemize{
\item The EQUALS operator is available for every attribute except
INSTANCE\\_LABELS.
\item The CONTAINS operator is available for the INSTANCE\\_LABELS and
MODEL attributes.
\item The IN and NOT\\_IN operators are available for the ARN,
OS\\_VERSION, MODEL, MANUFACTURER, and INSTANCE\\_ARN attributes.
\item The LESS\\_THAN, GREATER\\_THAN, LESS\\_THAN\\_OR\\_EQUALS, and
GREATER\\_THAN\\_OR\\_EQUALS operators are also available for the
OS\\_VERSION attribute.
}
\item Values: An array of one or more filter values.
\itemize{
\item The IN and NOT\\_IN operators take a values array that has one or
more elements.
\item The other operators require an array with a single element.
\item In a request, the AVAILABILITY attribute takes the following
values: AVAILABLE, HIGHLY\\_AVAILABLE, BUSY, or
TEMPORARY\\_NOT\\_AVAILABLE.
}
}}
}
\description{
Gets information about unique device types.
}
\section{Request syntax}{
\preformatted{svc$list_devices(
  arn = "string",
  nextToken = "string",
  filters = list(
    list(
      attribute = "ARN"|"PLATFORM"|"OS_VERSION"|"MODEL"|"AVAILABILITY"|"FORM_FACTOR"|"MANUFACTURER"|"REMOTE_ACCESS_ENABLED"|"REMOTE_DEBUG_ENABLED"|"INSTANCE_ARN"|"INSTANCE_LABELS"|"FLEET_TYPE",
      operator = "EQUALS"|"LESS_THAN"|"LESS_THAN_OR_EQUALS"|"GREATER_THAN"|"GREATER_THAN_OR_EQUALS"|"IN"|"NOT_IN"|"CONTAINS",
      values = list(
        "string"
      )
    )
  )
)
}
}

\examples{
# The following example returns information about the available devices in
# a specific project.
\donttest{svc$list_devices(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)}

}
\keyword{internal}
