% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_jobs}
\alias{devicefarm_list_jobs}
\title{Gets information about jobs for a given test run}
\usage{
devicefarm_list_jobs(arn, nextToken)
}
\arguments{
\item{arn}{[required] The run\'s Amazon Resource Name (ARN).}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Gets information about jobs for a given test run.
}
\section{Request syntax}{
\preformatted{svc$list_jobs(
  arn = "string",
  nextToken = "string"
)
}
}

\examples{
# The following example returns information about jobs in a specific
# project.
\dontrun{svc$list_jobs(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)}

}
\keyword{internal}
