% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_runs}
\alias{devicefarm_list_runs}
\title{Gets information about runs, given an AWS Device Farm project ARN}
\usage{
devicefarm_list_runs(arn, nextToken)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the project for which you want to list
runs.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Gets information about runs, given an AWS Device Farm project ARN.
}
\section{Request syntax}{
\preformatted{svc$list_runs(
  arn = "string",
  nextToken = "string"
)
}
}

\examples{
\dontrun{
# The following example returns information about a specific test run.
svc$list_runs(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:run:5e01a8c7-c861-4c0a-b1d5-5ec6e6c6dd23/...",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)
}

}
\keyword{internal}
