% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_samples}
\alias{devicefarm_list_samples}
\title{Gets information about samples, given an AWS Device Farm job ARN}
\usage{
devicefarm_list_samples(arn, nextToken)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the job used to list samples.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Gets information about samples, given an AWS Device Farm job ARN.
}
\section{Request syntax}{
\preformatted{devicefarm$list_samples(
  arn = "string",
  nextToken = "string"
)
}
}

\examples{
# The following example returns information about samples, given a
# specific Device Farm project.
\donttest{devicefarm$list_samples(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)}

}
\keyword{internal}
