% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_suites}
\alias{devicefarm_list_suites}
\title{Gets information about test suites for a given job}
\usage{
devicefarm_list_suites(arn, nextToken)
}
\arguments{
\item{arn}{[required] The job\'s Amazon Resource Name (ARN).}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Gets information about test suites for a given job.
}
\section{Request syntax}{
\preformatted{svc$list_suites(
  arn = "string",
  nextToken = "string"
)
}
}

\examples{
\dontrun{
# The following example returns information about suites, given a specific
# Device Farm job.
svc$list_suites(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:job:EXAMPLE-GUID-123-456",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)
}

}
\keyword{internal}
