% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_test_grid_session_actions}
\alias{devicefarm_list_test_grid_session_actions}
\title{Returns a list of the actions taken in a TestGridSession}
\usage{
devicefarm_list_test_grid_session_actions(sessionArn, maxResult,
  nextToken)
}
\arguments{
\item{sessionArn}{[required] The ARN of the session to retrieve.}

\item{maxResult}{The maximum number of sessions to return per response.}

\item{nextToken}{Pagination token.}
}
\description{
Returns a list of the actions taken in a TestGridSession.
}
\section{Request syntax}{
\preformatted{svc$list_test_grid_session_actions(
  sessionArn = "string",
  maxResult = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
