% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_unique_problems}
\alias{devicefarm_list_unique_problems}
\title{Gets information about unique problems}
\usage{
devicefarm_list_unique_problems(arn, nextToken)
}
\arguments{
\item{arn}{[required] The unique problems' ARNs.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Gets information about unique problems.
}
\section{Request syntax}{
\preformatted{svc$list_unique_problems(
  arn = "string",
  nextToken = "string"
)
}
}

\examples{
# The following example returns information about unique problems, given a
# specific Device Farm project.
\donttest{svc$list_unique_problems(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)}

}
\keyword{internal}
