% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_vpce_configurations}
\alias{devicefarm_list_vpce_configurations}
\title{Returns information about all Amazon Virtual Private Cloud (VPC)
endpoint configurations in the AWS account}
\usage{
devicefarm_list_vpce_configurations(maxResults, nextToken)
}
\arguments{
\item{maxResults}{An integer that specifies the maximum number of items you want to return
in the API response.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Returns information about all Amazon Virtual Private Cloud (VPC)
endpoint configurations in the AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_vpce_configurations(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
