% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_purchase_offering}
\alias{devicefarm_purchase_offering}
\title{Immediately purchases offerings for an AWS account}
\usage{
devicefarm_purchase_offering(offeringId, quantity, offeringPromotionId)
}
\arguments{
\item{offeringId}{The ID of the offering.}

\item{quantity}{The number of device slots you wish to purchase in an offering request.}

\item{offeringPromotionId}{The ID of the offering promotion to be applied to the purchase.}
}
\description{
Immediately purchases offerings for an AWS account. Offerings renew with
the latest total purchased quantity for an offering, unless the renewal
was overridden. The API returns a \code{NotEligible} error if the user is not
permitted to invoke the operation. Please contact
\href{mailto:aws-devicefarm-support@amazon.com}{aws-devicefarm-support@amazon.com} if you believe that you should be
able to invoke this operation.
}
\section{Request syntax}{
\preformatted{svc$purchase_offering(
  offeringId = "string",
  quantity = 123,
  offeringPromotionId = "string"
)
}
}

\examples{
# The following example purchases a specific device slot offering.
\donttest{svc$purchase_offering(
  offeringId = "D68B3C05-1BA6-4360-BC69-12345EXAMPLE",
  quantity = 1L
)}

}
\keyword{internal}
