% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_untag_resource}
\alias{devicefarm_untag_resource}
\title{Deletes the specified tags from a resource}
\usage{
devicefarm_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource or resources from which
to delete tags. You can associate tags with the following Device Farm
resources: \code{PROJECT}, \code{RUN}, \code{NETWORK_PROFILE}, \code{INSTANCE_PROFILE},
\code{DEVICE_INSTANCE}, \code{SESSION}, \code{DEVICE_POOL}, \code{DEVICE}, and
\code{VPCE_CONFIGURATION}.}

\item{TagKeys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes the specified tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
