% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_update_device_pool}
\alias{devicefarm_update_device_pool}
\title{Modifies the name, description, and rules in a device pool given the
attributes and the pool ARN}
\usage{
devicefarm_update_device_pool(arn, name, description, rules, maxDevices,
  clearMaxDevices)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the Device Farm device pool to update.}

\item{name}{A string that represents the name of the device pool to update.}

\item{description}{A description of the device pool to update.}

\item{rules}{Represents the rules to modify for the device pool. Updating rules is
optional. If you update rules for your request, the update replaces the
existing rules.}

\item{maxDevices}{The number of devices that Device Farm can add to your device pool.
Device Farm adds devices that are available and that meet the criteria
that you assign for the \code{rules} parameter. Depending on how many devices
meet these constraints, your device pool might contain fewer devices
than the value for this parameter.

By specifying the maximum number of devices, you can control the costs
that you incur by running tests.

If you use this parameter in your request, you cannot use the
\code{clearMaxDevices} parameter in the same request.}

\item{clearMaxDevices}{Sets whether the \code{maxDevices} parameter applies to your device pool. If
you set this parameter to \code{true}, the \code{maxDevices} parameter does not
apply, and Device Farm does not limit the number of devices that it adds
to your device pool. In this case, Device Farm adds all available
devices that meet the criteria specified in the \code{rules} parameter.

If you use this parameter in your request, you cannot use the
\code{maxDevices} parameter in the same request.}
}
\description{
Modifies the name, description, and rules in a device pool given the
attributes and the pool ARN. Rule updates are all-or-nothing, meaning
they can only be updated as a whole (or not at all).
}
\section{Request syntax}{
\preformatted{svc$update_device_pool(
  arn = "string",
  name = "string",
  description = "string",
  rules = list(
    list(
      attribute = "ARN"|"PLATFORM"|"FORM_FACTOR"|"MANUFACTURER"|"REMOTE_ACCESS_ENABLED"|"REMOTE_DEBUG_ENABLED"|"APPIUM_VERSION"|"INSTANCE_ARN"|"INSTANCE_LABELS"|"FLEET_TYPE"|"OS_VERSION"|"MODEL"|"AVAILABILITY",
      operator = "EQUALS"|"LESS_THAN"|"LESS_THAN_OR_EQUALS"|"GREATER_THAN"|"GREATER_THAN_OR_EQUALS"|"IN"|"NOT_IN"|"CONTAINS",
      value = "string"
    )
  ),
  maxDevices = 123,
  clearMaxDevices = TRUE|FALSE
)
}
}

\examples{
\dontrun{
# The following example updates the specified device pool with a new name
# and description. It also enables remote access of devices in the device
# pool.
svc$update_device_pool(
  name = "NewName",
  arn = "arn:aws:devicefarm:us-west-2::devicepool:082d10e5-d7d7-48a5-ba5c-12345EXAMPLE",
  description = "NewDescription",
  rules = list(
    list(
      value = "True",
      attribute = "REMOTE_ACCESS_ENABLED",
      operator = "EQUALS"
    )
  )
)
}

}
\keyword{internal}
