% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_add_notification_channel}
\alias{devopsguru_add_notification_channel}
\title{Adds a notification channel to DevOps Guru}
\usage{
devopsguru_add_notification_channel(Config)
}
\arguments{
\item{Config}{[required] A \code{NotificationChannelConfig} object that specifies what type of
notification channel to add. The one supported notification channel is
Amazon Simple Notification Service (Amazon SNS).}
}
\description{
Adds a notification channel to DevOps Guru. A notification channel is
used to notify you about important DevOps Guru events, such as when an
insight is generated.

If you use an Amazon SNS topic in another account, you must attach a
policy to it that grants DevOps Guru permission to it notifications.
DevOps Guru adds the required policy on your behalf to send
notifications using Amazon SNS in your account. For more information,
see \href{https://docs.aws.amazon.com/devops-guru/latest/userguide/sns-required-permissions.html}{Permissions for cross account Amazon SNS topics}.

If you use an Amazon SNS topic that is encrypted by an AWS Key
Management Service customer-managed key (CMK), then you must add
permissions to the CMK. For more information, see \href{https://docs.aws.amazon.com/devops-guru/latest/userguide/sns-kms-permissions.html}{Permissions for AWS KMS–encrypted Amazon SNS topics}.
}
\section{Request syntax}{
\preformatted{svc$add_notification_channel(
  Config = list(
    Sns = list(
      TopicArn = "string"
    )
  )
)
}
}

\keyword{internal}
