% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_describe_account_health}
\alias{devopsguru_describe_account_health}
\title{Returns the number of open reactive insights, the number of open
proactive insights, and the number of metrics analyzed in your AWS
account}
\usage{
devopsguru_describe_account_health()
}
\description{
Returns the number of open reactive insights, the number of open
proactive insights, and the number of metrics analyzed in your AWS
account. Use these numbers to gauge the health of operations in your AWS
account.
}
\section{Request syntax}{
\preformatted{svc$describe_account_health()
}
}

\keyword{internal}
