% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_describe_anomaly}
\alias{devopsguru_describe_anomaly}
\title{Returns details about an anomaly that you specify using its ID}
\usage{
devopsguru_describe_anomaly(Id)
}
\arguments{
\item{Id}{[required] The ID of the anomaly.}
}
\description{
Returns details about an anomaly that you specify using its ID.
}
\section{Request syntax}{
\preformatted{svc$describe_anomaly(
  Id = "string"
)
}
}

\keyword{internal}
