% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_describe_resource_collection_health}
\alias{devopsguru_describe_resource_collection_health}
\title{Returns the number of open proactive insights, open reactive insights,
and the Mean Time to Recover (MTTR) for all closed insights in resource
collections in your account}
\usage{
devopsguru_describe_resource_collection_health(ResourceCollectionType,
  NextToken)
}
\arguments{
\item{ResourceCollectionType}{[required] An AWS resource collection type. This type specifies how analyzed AWS
resources are defined. The one type of AWS resource collection supported
is AWS CloudFormation stacks. DevOps Guru can be configured to analyze
only the AWS resources that are defined in the stacks.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Returns the number of open proactive insights, open reactive insights,
and the Mean Time to Recover (MTTR) for all closed insights in resource
collections in your account. You specify the type of AWS resources
collection. The one type of AWS resource collection supported is AWS
CloudFormation stacks. DevOps Guru can be configured to analyze only the
AWS resources that are defined in the stacks.
}
\section{Request syntax}{
\preformatted{svc$describe_resource_collection_health(
  ResourceCollectionType = "AWS_CLOUD_FORMATION",
  NextToken = "string"
)
}
}

\keyword{internal}
