% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_describe_service_integration}
\alias{devopsguru_describe_service_integration}
\title{Returns the integration status of services that are integrated with
DevOps Guru}
\usage{
devopsguru_describe_service_integration()
}
\description{
Returns the integration status of services that are integrated with
DevOps Guru. The one service that can be integrated with DevOps Guru is
AWS Systems Manager, which can be used to create an OpsItem for each
generated insight.
}
\section{Request syntax}{
\preformatted{svc$describe_service_integration()
}
}

\keyword{internal}
