% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_get_resource_collection}
\alias{devopsguru_get_resource_collection}
\title{Returns lists AWS resources that are of the specified resource
collection type}
\usage{
devopsguru_get_resource_collection(ResourceCollectionType, NextToken)
}
\arguments{
\item{ResourceCollectionType}{[required] The type of AWS resource collections to return. The one valid value is
\code{CLOUD_FORMATION} for AWS CloudFormation stacks.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Returns lists AWS resources that are of the specified resource
collection type. The one type of AWS resource collection supported is
AWS CloudFormation stacks. DevOps Guru can be configured to analyze only
the AWS resources that are defined in the stacks.
}
\section{Request syntax}{
\preformatted{svc$get_resource_collection(
  ResourceCollectionType = "AWS_CLOUD_FORMATION",
  NextToken = "string"
)
}
}

\keyword{internal}
