% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_list_anomalies_for_insight}
\alias{devopsguru_list_anomalies_for_insight}
\title{Returns a list of the anomalies that belong to an insight that you
specify using its ID}
\usage{
devopsguru_list_anomalies_for_insight(InsightId, StartTimeRange,
  MaxResults, NextToken)
}
\arguments{
\item{InsightId}{[required] The ID of the insight. The returned anomalies belong to this insight.}

\item{StartTimeRange}{A time range used to specify when the requested anomalies started. All
returned anomalies started during this time range.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Returns a list of the anomalies that belong to an insight that you
specify using its ID.
}
\section{Request syntax}{
\preformatted{svc$list_anomalies_for_insight(
  InsightId = "string",
  StartTimeRange = list(
    FromTime = as.POSIXct(
      "2015-01-01"
    ),
    ToTime = as.POSIXct(
      "2015-01-01"
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
