% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_list_events}
\alias{devopsguru_list_events}
\title{Returns a list of the events emitted by the resources that are evaluated
by DevOps Guru}
\usage{
devopsguru_list_events(Filters, MaxResults, NextToken)
}
\arguments{
\item{Filters}{[required] A \code{ListEventsFilters} object used to specify which events to return.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Returns a list of the events emitted by the resources that are evaluated
by DevOps Guru. You can use filters to specify which events are
returned.
}
\section{Request syntax}{
\preformatted{svc$list_events(
  Filters = list(
    InsightId = "string",
    EventTimeRange = list(
      FromTime = as.POSIXct(
        "2015-01-01"
      ),
      ToTime = as.POSIXct(
        "2015-01-01"
      )
    ),
    EventClass = "INFRASTRUCTURE"|"DEPLOYMENT"|"SECURITY_CHANGE"|"CONFIG_CHANGE"|"SCHEMA_CHANGE",
    EventSource = "string",
    DataSource = "AWS_CLOUD_TRAIL"|"AWS_CODE_DEPLOY",
    ResourceCollection = list(
      CloudFormation = list(
        StackNames = list(
          "string"
        )
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
