% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_list_recommendations}
\alias{devopsguru_list_recommendations}
\title{Returns a list of a specified insight's recommendations}
\usage{
devopsguru_list_recommendations(InsightId, NextToken)
}
\arguments{
\item{InsightId}{[required] The ID of the requested insight.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Returns a list of a specified insight's recommendations. Each
recommendation includes a list of related metrics and a list of related
events.
}
\section{Request syntax}{
\preformatted{svc$list_recommendations(
  InsightId = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
