% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_put_feedback}
\alias{devopsguru_put_feedback}
\title{Collects customer feedback about the specified insight}
\usage{
devopsguru_put_feedback(InsightFeedback)
}
\arguments{
\item{InsightFeedback}{The feedback from customers is about the recommendations in this
insight.}
}
\description{
Collects customer feedback about the specified insight.
}
\section{Request syntax}{
\preformatted{svc$put_feedback(
  InsightFeedback = list(
    Id = "string",
    Feedback = "VALID_COLLECTION"|"RECOMMENDATION_USEFUL"|"ALERT_TOO_SENSITIVE"|"DATA_NOISY_ANOMALY"|"DATA_INCORRECT"
  )
)
}
}

\keyword{internal}
