% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_search_insights}
\alias{devopsguru_search_insights}
\title{Returns a list of insights in your AWS account}
\usage{
devopsguru_search_insights(StartTimeRange, Filters, MaxResults,
  NextToken, Type)
}
\arguments{
\item{StartTimeRange}{[required] The start of the time range passed in. Returned insights occurred after
this time.}

\item{Filters}{A \code{SearchInsightsFilters} object that is used to set the severity and
status filters on your insight search.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}

\item{Type}{[required] The type of insights you are searching for (\code{REACTIVE} or \code{PROACTIVE}).}
}
\description{
Returns a list of insights in your AWS account. You can specify which
insights are returned by their start time, one or more statuses
(\code{ONGOING}, \code{CLOSED}, and \code{CLOSED}), one or more severities (\code{LOW},
\code{MEDIUM}, and \code{HIGH}), and type (\code{REACTIVE} or \code{PROACTIVE}).

Use the \code{Filters} parameter to specify status and severity search
parameters. Use the \code{Type} parameter to specify \code{REACTIVE} or
\code{PROACTIVE} in your search.
}
\section{Request syntax}{
\preformatted{svc$search_insights(
  StartTimeRange = list(
    FromTime = as.POSIXct(
      "2015-01-01"
    ),
    ToTime = as.POSIXct(
      "2015-01-01"
    )
  ),
  Filters = list(
    Severities = list(
      "LOW"|"MEDIUM"|"HIGH"
    ),
    Statuses = list(
      "ONGOING"|"CLOSED"
    ),
    ResourceCollection = list(
      CloudFormation = list(
        StackNames = list(
          "string"
        )
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string",
  Type = "REACTIVE"|"PROACTIVE"
)
}
}

\keyword{internal}
