% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_update_resource_collection}
\alias{devopsguru_update_resource_collection}
\title{Updates the collection of resources that DevOps Guru analyzes}
\usage{
devopsguru_update_resource_collection(Action, ResourceCollection)
}
\arguments{
\item{Action}{[required] Specifies if the resource collection in the request is added or deleted
to the resource collection.}

\item{ResourceCollection}{[required]}
}
\description{
Updates the collection of resources that DevOps Guru analyzes. The one
type of AWS resource collection supported is AWS CloudFormation stacks.
DevOps Guru can be configured to analyze only the AWS resources that are
defined in the stacks. This method also creates the IAM role required
for you to use DevOps Guru.
}
\section{Request syntax}{
\preformatted{svc$update_resource_collection(
  Action = "ADD"|"REMOVE",
  ResourceCollection = list(
    CloudFormation = list(
      StackNames = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
