% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_connections_on_interconnect}
\alias{directconnect_describe_connections_on_interconnect}
\title{Deprecated}
\usage{
directconnect_describe_connections_on_interconnect(interconnectId)
}
\arguments{
\item{interconnectId}{[required] The ID of the interconnect.}
}
\description{
Deprecated. Use DescribeHostedConnections instead.
}
\details{
Lists the connections that have been provisioned on the specified
interconnect.

Intended for use by AWS Direct Connect Partners only.
}
\section{Request syntax}{
\preformatted{directconnect$describe_connections_on_interconnect(
  interconnectId = "string"
)
}
}

\keyword{internal}
