% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_direct_connect_gateways}
\alias{directconnect_describe_direct_connect_gateways}
\title{Lists all your Direct Connect gateways or only the specified Direct
Connect gateway}
\usage{
directconnect_describe_direct_connect_gateways(directConnectGatewayId,
  maxResults, nextToken)
}
\arguments{
\item{directConnectGatewayId}{The ID of the Direct Connect gateway.}

\item{maxResults}{The maximum number of Direct Connect gateways to return per page.}

\item{nextToken}{The token provided in the previous call to retrieve the next page.}
}
\description{
Lists all your Direct Connect gateways or only the specified Direct
Connect gateway. Deleted Direct Connect gateways are not returned.
}
\section{Request syntax}{
\preformatted{directconnect$describe_direct_connect_gateways(
  directConnectGatewayId = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
