% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_deregister_certificate}
\alias{directoryservice_deregister_certificate}
\title{Deletes from the system the certificate that was registered for a
secured LDAP connection}
\usage{
directoryservice_deregister_certificate(DirectoryId, CertificateId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{CertificateId}{[required] The identifier of the certificate.}
}
\description{
Deletes from the system the certificate that was registered for a
secured LDAP connection.
}
\section{Request syntax}{
\preformatted{svc$deregister_certificate(
  DirectoryId = "string",
  CertificateId = "string"
)
}
}

\keyword{internal}
