% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_disable_ldaps}
\alias{directoryservice_disable_ldaps}
\title{Deactivates LDAP secure calls for the specified directory}
\usage{
directoryservice_disable_ldaps(DirectoryId, Type)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{Type}{The type of LDAP security that the customer wants to enable. The
security can be either server or client, but currently only the default
\code{Client} is supported.}
}
\description{
Deactivates LDAP secure calls for the specified directory.
}
\section{Request syntax}{
\preformatted{svc$disable_ldaps(
  DirectoryId = "string",
  Type = "Client"
)
}
}

\keyword{internal}
