% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_register_certificate}
\alias{directoryservice_register_certificate}
\title{Registers a certificate for secured LDAP connection}
\usage{
directoryservice_register_certificate(DirectoryId, CertificateData)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{CertificateData}{[required] The certificate PEM string that needs to be registered.}
}
\description{
Registers a certificate for secured LDAP connection.
}
\section{Request syntax}{
\preformatted{svc$register_certificate(
  DirectoryId = "string",
  CertificateData = "string"
)
}
}

\keyword{internal}
