% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_delete_backup}
\alias{dynamodb_delete_backup}
\title{Deletes an existing backup of a table}
\usage{
dynamodb_delete_backup(BackupArn)
}
\arguments{
\item{BackupArn}{[required] The ARN associated with the backup.}
}
\description{
Deletes an existing backup of a table.

You can call \code{\link[=dynamodb_delete_backup]{delete_backup}} at a maximum rate
of 10 times per second.
}
\section{Request syntax}{
\preformatted{svc$delete_backup(
  BackupArn = "string"
)
}
}

\keyword{internal}
