% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_endpoints}
\alias{dynamodb_describe_endpoints}
\title{Returns the regional endpoint information}
\usage{
dynamodb_describe_endpoints()
}
\description{
Returns the regional endpoint information.
}
\section{Request syntax}{
\preformatted{dynamodb$describe_endpoints()
}
}

\keyword{internal}
