% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_global_table}
\alias{dynamodb_describe_global_table}
\title{Returns information about the specified global table}
\usage{
dynamodb_describe_global_table(GlobalTableName)
}
\arguments{
\item{GlobalTableName}{[required] The name of the global table.}
}
\description{
Returns information about the specified global table.

This operation only applies to \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/globaltables.V1.html}{Version 2017.11.29}
of global tables. If you are using global tables \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/globaltables.V2.html}{Version 2019.11.21}
you can use
\href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_DescribeTable.html}{DescribeTable}
instead.
}
\section{Request syntax}{
\preformatted{svc$describe_global_table(
  GlobalTableName = "string"
)
}
}

\keyword{internal}
