% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_list_exports}
\alias{dynamodb_list_exports}
\title{Lists completed exports within the past 90 days}
\usage{
dynamodb_list_exports(TableArn, MaxResults, NextToken)
}
\arguments{
\item{TableArn}{The Amazon Resource Name (ARN) associated with the exported table.}

\item{MaxResults}{Maximum number of results to return per page.}

\item{NextToken}{An optional string that, if supplied, must be copied from the output of
a previous call to \code{\link[=dynamodb_list_exports]{list_exports}}. When
provided in this manner, the API fetches the next page of results.}
}
\description{
Lists completed exports within the past 90 days.
}
\section{Request syntax}{
\preformatted{svc$list_exports(
  TableArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
