% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_transact_get_items}
\alias{dynamodb_transact_get_items}
\title{TransactGetItems is a synchronous operation that atomically retrieves
multiple items from one or more tables (but not from indexes) in a
single account and region}
\usage{
dynamodb_transact_get_items(TransactItems, ReturnConsumedCapacity)
}
\arguments{
\item{TransactItems}{[required] An ordered array of up to 10 \code{TransactGetItem} objects, each of which
contains a \code{Get} structure.}

\item{ReturnConsumedCapacity}{A value of \code{TOTAL} causes consumed capacity information to be returned,
and a value of \code{NONE} prevents that information from being returned. No
other value is valid.}
}
\description{
\code{TransactGetItems} is a synchronous operation that atomically retrieves
multiple items from one or more tables (but not from indexes) in a
single account and region. A \code{TransactGetItems} call can contain up to
10 \code{TransactGetItem} objects, each of which contains a \code{Get} structure
that specifies an item to retrieve from a table in the account and
region. A call to \code{TransactGetItems} cannot retrieve items from tables
in more than one AWS account or region.
}
\details{
DynamoDB rejects the entire \code{TransactGetItems} request if any of the
following is true:
\itemize{
\item A conflicting operation is in the process of updating an item to be
read.
\item There is insufficient provisioned capacity for the transaction to be
completed.
\item There is a user error, such as an invalid data format.
}
}
\section{Request syntax}{
\preformatted{svc$transact_get_items(
  TransactItems = list(
    list(
      Get = list(
        Key = list(
          list(
            S = "string",
            N = "string",
            B = raw,
            SS = list(
              "string"
            ),
            NS = list(
              "string"
            ),
            BS = list(
              raw
            ),
            M = list(
              list()
            ),
            L = list(
              list()
            ),
            NULL = TRUE|FALSE,
            BOOL = TRUE|FALSE
          )
        ),
        TableName = "string",
        ProjectionExpression = "string",
        ExpressionAttributeNames = list(
          "string"
        )
      )
    )
  ),
  ReturnConsumedCapacity = "INDEXES"|"TOTAL"|"NONE"
)
}
}

\keyword{internal}
