% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_update_contributor_insights}
\alias{dynamodb_update_contributor_insights}
\title{Updates the status for contributor insights for a specific table or
index}
\usage{
dynamodb_update_contributor_insights(TableName, IndexName,
  ContributorInsightsAction)
}
\arguments{
\item{TableName}{[required] The name of the table.}

\item{IndexName}{The global secondary index name, if applicable.}

\item{ContributorInsightsAction}{[required] Represents the contributor insights action.}
}
\description{
Updates the status for contributor insights for a specific table or
index.
}
\section{Request syntax}{
\preformatted{svc$update_contributor_insights(
  TableName = "string",
  IndexName = "string",
  ContributorInsightsAction = "ENABLE"|"DISABLE"
)
}
}

\keyword{internal}
