% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebs_operations.R
\name{ebs_get_snapshot_block}
\alias{ebs_get_snapshot_block}
\title{Returns the data in a block in an Amazon Elastic Block Store snapshot}
\usage{
ebs_get_snapshot_block(SnapshotId, BlockIndex, BlockToken)
}
\arguments{
\item{SnapshotId}{[required] The ID of the snapshot containing the block from which to get data.}

\item{BlockIndex}{[required] The block index of the block from which to get data.

Obtain the \verb{block index} by running the \verb{list changed blocks} or
\verb{list snapshot blocks} operations.}

\item{BlockToken}{[required] The block token of the block from which to get data.

Obtain the \verb{block token} by running the \verb{list changed blocks} or
\verb{list snapshot blocks} operations.}
}
\description{
Returns the data in a block in an Amazon Elastic Block Store snapshot.
}
\section{Request syntax}{
\preformatted{svc$get_snapshot_block(
  SnapshotId = "string",
  BlockIndex = 123,
  BlockToken = "string"
)
}
}

\keyword{internal}
