% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebs_operations.R
\name{ebs_list_changed_blocks}
\alias{ebs_list_changed_blocks}
\title{Returns information about the blocks that are different between two
Amazon Elastic Block Store snapshots of the same volume/snapshot lineage}
\usage{
ebs_list_changed_blocks(FirstSnapshotId, SecondSnapshotId, NextToken,
  MaxResults, StartingBlockIndex)
}
\arguments{
\item{FirstSnapshotId}{The ID of the first snapshot to use for the comparison.

The \code{FirstSnapshotID} parameter must be specified with a
\code{SecondSnapshotId} parameter; otherwise, an error occurs.}

\item{SecondSnapshotId}{[required] The ID of the second snapshot to use for the comparison.

The \code{SecondSnapshotId} parameter must be specified with a
\code{FirstSnapshotID} parameter; otherwise, an error occurs.}

\item{NextToken}{The token to request the next page of results.}

\item{MaxResults}{The number of results to return.}

\item{StartingBlockIndex}{The block index from which the comparison should start.

The list in the response will start from this block index or the next
valid block index in the snapshots.}
}
\description{
Returns information about the blocks that are different between two
Amazon Elastic Block Store snapshots of the same volume/snapshot
lineage.
}
\section{Request syntax}{
\preformatted{svc$list_changed_blocks(
  FirstSnapshotId = "string",
  SecondSnapshotId = "string",
  NextToken = "string",
  MaxResults = 123,
  StartingBlockIndex = 123
)
}
}

\keyword{internal}
