% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebs_operations.R
\name{ebs_list_snapshot_blocks}
\alias{ebs_list_snapshot_blocks}
\title{Returns information about the blocks in an Amazon Elastic Block Store
snapshot}
\usage{
ebs_list_snapshot_blocks(SnapshotId, NextToken, MaxResults,
  StartingBlockIndex)
}
\arguments{
\item{SnapshotId}{[required] The ID of the snapshot from which to get block indexes and block tokens.}

\item{NextToken}{The token to request the next page of results.}

\item{MaxResults}{The number of results to return.}

\item{StartingBlockIndex}{The block index from which the list should start. The list in the
response will start from this block index or the next valid block index
in the snapshot.}
}
\description{
Returns information about the blocks in an Amazon Elastic Block Store
snapshot.
}
\section{Request syntax}{
\preformatted{svc$list_snapshot_blocks(
  SnapshotId = "string",
  NextToken = "string",
  MaxResults = 123,
  StartingBlockIndex = 123
)
}
}

\keyword{internal}
