% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_cancel_reserved_instances_listing}
\alias{ec2_cancel_reserved_instances_listing}
\title{Cancels the specified Reserved Instance listing in the Reserved Instance
Marketplace}
\usage{
ec2_cancel_reserved_instances_listing(ReservedInstancesListingId)
}
\arguments{
\item{ReservedInstancesListingId}{[required] The ID of the Reserved Instance listing.}
}
\description{
Cancels the specified Reserved Instance listing in the Reserved Instance
Marketplace.
}
\details{
For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-market-general.html}{Reserved Instance Marketplace}
in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\section{Request syntax}{
\preformatted{ec2$cancel_reserved_instances_listing(
  ReservedInstancesListingId = "string"
)
}
}

\keyword{internal}
