% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_route_table}
\alias{ec2_delete_route_table}
\title{Deletes the specified route table}
\usage{
ec2_delete_route_table(DryRun, RouteTableId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{RouteTableId}{[required] The ID of the route table.}
}
\description{
Deletes the specified route table. You must disassociate the route table
from any subnets before you can delete it. You can\'t delete the main
route table.
}
\section{Request syntax}{
\preformatted{svc$delete_route_table(
  DryRun = TRUE|FALSE,
  RouteTableId = "string"
)
}
}

\examples{
\dontrun{
# This example deletes the specified route table.
svc$delete_route_table(
  RouteTableId = "rtb-22574640"
)
}

}
\keyword{internal}
