% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_vpn_tunnel_certificate}
\alias{ec2_modify_vpn_tunnel_certificate}
\title{Modifies the VPN tunnel endpoint certificate}
\usage{
ec2_modify_vpn_tunnel_certificate(VpnConnectionId,
  VpnTunnelOutsideIpAddress, DryRun)
}
\arguments{
\item{VpnConnectionId}{[required] The ID of the AWS Site-to-Site VPN connection.}

\item{VpnTunnelOutsideIpAddress}{[required] The external IP address of the VPN tunnel.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Modifies the VPN tunnel endpoint certificate.
}
\section{Request syntax}{
\preformatted{svc$modify_vpn_tunnel_certificate(
  VpnConnectionId = "string",
  VpnTunnelOutsideIpAddress = "string",
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
