% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_delete_registry_policy}
\alias{ecr_delete_registry_policy}
\title{Deletes the registry permissions policy}
\usage{
ecr_delete_registry_policy()
}
\description{
Deletes the registry permissions policy.
}
\section{Request syntax}{
\preformatted{svc$delete_registry_policy()
}
}

\keyword{internal}
