% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_describe_registry}
\alias{ecr_describe_registry}
\title{Describes the settings for a registry}
\usage{
ecr_describe_registry()
}
\description{
Describes the settings for a registry. The replication configuration for
a repository can be created or updated with the
\code{\link[=ecr_put_replication_configuration]{put_replication_configuration}} API
action.
}
\section{Request syntax}{
\preformatted{svc$describe_registry()
}
}

\keyword{internal}
