% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_delete_repository_policy}
\alias{ecrpublic_delete_repository_policy}
\title{Deletes the repository policy associated with the specified repository}
\usage{
ecrpublic_delete_repository_policy(registryId, repositoryName)
}
\arguments{
\item{registryId}{The AWS account ID associated with the public registry that contains the
repository policy to delete. If you do not specify a registry, the
default public registry is assumed.}

\item{repositoryName}{[required] The name of the repository that is associated with the repository policy
to delete.}
}
\description{
Deletes the repository policy associated with the specified repository.
}
\section{Request syntax}{
\preformatted{svc$delete_repository_policy(
  registryId = "string",
  repositoryName = "string"
)
}
}

\keyword{internal}
