% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_get_registry_catalog_data}
\alias{ecrpublic_get_registry_catalog_data}
\title{Retrieves catalog metadata for a public registry}
\usage{
ecrpublic_get_registry_catalog_data()
}
\description{
Retrieves catalog metadata for a public registry.
}
\section{Request syntax}{
\preformatted{svc$get_registry_catalog_data()
}
}

\keyword{internal}
