% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_get_repository_catalog_data}
\alias{ecrpublic_get_repository_catalog_data}
\title{Retrieve catalog metadata for a repository in a public registry}
\usage{
ecrpublic_get_repository_catalog_data(registryId, repositoryName)
}
\arguments{
\item{registryId}{The AWS account ID associated with the registry that contains the
repositories to be described. If you do not specify a registry, the
default public registry is assumed.}

\item{repositoryName}{[required] The name of the repository to retrieve the catalog metadata for.}
}
\description{
Retrieve catalog metadata for a repository in a public registry. This
metadata is displayed publicly in the Amazon ECR Public Gallery.
}
\section{Request syntax}{
\preformatted{svc$get_repository_catalog_data(
  registryId = "string",
  repositoryName = "string"
)
}
}

\keyword{internal}
