% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_get_repository_policy}
\alias{ecrpublic_get_repository_policy}
\title{Retrieves the repository policy for the specified repository}
\usage{
ecrpublic_get_repository_policy(registryId, repositoryName)
}
\arguments{
\item{registryId}{The AWS account ID associated with the public registry that contains the
repository. If you do not specify a registry, the default public
registry is assumed.}

\item{repositoryName}{[required] The name of the repository with the policy to retrieve.}
}
\description{
Retrieves the repository policy for the specified repository.
}
\section{Request syntax}{
\preformatted{svc$get_repository_policy(
  registryId = "string",
  repositoryName = "string"
)
}
}

\keyword{internal}
