% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_fargate_profile}
\alias{eks_describe_fargate_profile}
\title{Returns descriptive information about an AWS Fargate profile}
\usage{
eks_describe_fargate_profile(clusterName, fargateProfileName)
}
\arguments{
\item{clusterName}{[required] The name of the Amazon EKS cluster associated with the Fargate profile.}

\item{fargateProfileName}{[required] The name of the Fargate profile to describe.}
}
\description{
Returns descriptive information about an AWS Fargate profile.
}
\section{Request syntax}{
\preformatted{svc$describe_fargate_profile(
  clusterName = "string",
  fargateProfileName = "string"
)
}
}

\keyword{internal}
