% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_batch_apply_update_action}
\alias{elasticache_batch_apply_update_action}
\title{Apply the service update}
\usage{
elasticache_batch_apply_update_action(ReplicationGroupIds,
  ServiceUpdateName)
}
\arguments{
\item{ReplicationGroupIds}{[required] The replication group IDs}

\item{ServiceUpdateName}{[required] The unique ID of the service update}
}
\description{
Apply the service update. For more information on service updates and
applying them, see Applying Service Updates.
}
\section{Request syntax}{
\preformatted{svc$batch_apply_update_action(
  ReplicationGroupIds = list(
    "string"
  ),
  ServiceUpdateName = "string"
)
}
}

\keyword{internal}
