% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_create_replication_group}
\alias{elasticache_create_replication_group}
\title{Creates a Redis (cluster mode disabled) or a Redis (cluster mode
enabled) replication group}
\usage{
elasticache_create_replication_group(ReplicationGroupId,
  ReplicationGroupDescription, PrimaryClusterId, AutomaticFailoverEnabled,
  NumCacheClusters, PreferredCacheClusterAZs, NumNodeGroups,
  ReplicasPerNodeGroup, NodeGroupConfiguration, CacheNodeType, Engine,
  EngineVersion, CacheParameterGroupName, CacheSubnetGroupName,
  CacheSecurityGroupNames, SecurityGroupIds, Tags, SnapshotArns,
  SnapshotName, PreferredMaintenanceWindow, Port, NotificationTopicArn,
  AutoMinorVersionUpgrade, SnapshotRetentionLimit, SnapshotWindow,
  AuthToken, TransitEncryptionEnabled, AtRestEncryptionEnabled)
}
\arguments{
\item{ReplicationGroupId}{[required] The replication group identifier. This parameter is stored as a
lowercase string.

Constraints:
\itemize{
\item A name must contain from 1 to 20 alphanumeric characters or hyphens.
\item The first character must be a letter.
\item A name cannot end with a hyphen or contain two consecutive hyphens.
}}

\item{ReplicationGroupDescription}{[required] A user-created description for the replication group.}

\item{PrimaryClusterId}{The identifier of the cluster that serves as the primary for this
replication group. This cluster must already exist and have a status of
\code{available}.

This parameter is not required if \code{NumCacheClusters}, \code{NumNodeGroups},
or \code{ReplicasPerNodeGroup} is specified.}

\item{AutomaticFailoverEnabled}{Specifies whether a read-only replica is automatically promoted to
read/write primary if the existing primary fails.

If \code{true}, Multi-AZ is enabled for this replication group. If \code{false},
Multi-AZ is disabled for this replication group.

\code{AutomaticFailoverEnabled} must be enabled for Redis (cluster mode
enabled) replication groups.

Default: false

Amazon ElastiCache for Redis does not support Multi-AZ with automatic
failover on:
\itemize{
\item Redis versions earlier than 2.8.6.
\item Redis (cluster mode disabled): T1 and T2 cache node types.
\item Redis (cluster mode enabled): T1 node types.
}}

\item{NumCacheClusters}{The number of clusters this replication group initially has.

This parameter is not used if there is more than one node group (shard).
You should use \code{ReplicasPerNodeGroup} instead.

If \code{AutomaticFailoverEnabled} is \code{true}, the value of this parameter
must be at least 2. If \code{AutomaticFailoverEnabled} is \code{false} you can
omit this parameter (it will default to 1), or you can explicitly set it
to a value between 2 and 6.

The maximum permitted value for \code{NumCacheClusters} is 6 (1 primary plus
5 replicas).}

\item{PreferredCacheClusterAZs}{A list of EC2 Availability Zones in which the replication group's
clusters are created. The order of the Availability Zones in the list is
the order in which clusters are allocated. The primary cluster is
created in the first AZ in the list.

This parameter is not used if there is more than one node group (shard).
You should use \code{NodeGroupConfiguration} instead.

If you are creating your replication group in an Amazon VPC
(recommended), you can only locate clusters in Availability Zones
associated with the subnets in the selected subnet group.

The number of Availability Zones listed must equal the value of
\code{NumCacheClusters}.

Default: system chosen Availability Zones.}

\item{NumNodeGroups}{An optional parameter that specifies the number of node groups (shards)
for this Redis (cluster mode enabled) replication group. For Redis
(cluster mode disabled) either omit this parameter or set it to 1.

Default: 1}

\item{ReplicasPerNodeGroup}{An optional parameter that specifies the number of replica nodes in each
node group (shard). Valid values are 0 to 5.}

\item{NodeGroupConfiguration}{A list of node group (shard) configuration options. Each node group
(shard) configuration has the following members:
\code{PrimaryAvailabilityZone}, \code{ReplicaAvailabilityZones}, \code{ReplicaCount},
and \code{Slots}.

If you're creating a Redis (cluster mode disabled) or a Redis (cluster
mode enabled) replication group, you can use this parameter to
individually configure each node group (shard), or you can omit this
parameter. However, when seeding a Redis (cluster mode enabled) cluster
from a S3 rdb file, you must configure each node group (shard) using
this parameter because you must specify the slots for each node group.}

\item{CacheNodeType}{The compute and memory capacity of the nodes in the node group (shard).

The following node types are supported by ElastiCache. Generally
speaking, the current generation types provide more memory and
computational power at lower cost when compared to their equivalent
previous generation counterparts.
\itemize{
\item General purpose:
\itemize{
\item Current generation:

\strong{T2 node types:} \code{cache.t2.micro}, \code{cache.t2.small},
\code{cache.t2.medium}

\strong{M3 node types:} \code{cache.m3.medium}, \code{cache.m3.large},
\code{cache.m3.xlarge}, \code{cache.m3.2xlarge}

\strong{M4 node types:} \code{cache.m4.large}, \code{cache.m4.xlarge},
\code{cache.m4.2xlarge}, \code{cache.m4.4xlarge}, \code{cache.m4.10xlarge}
\item Previous generation: (not recommended)

\strong{T1 node types:} \code{cache.t1.micro}

\strong{M1 node types:} \code{cache.m1.small}, \code{cache.m1.medium},
\code{cache.m1.large}, \code{cache.m1.xlarge}
}
\item Compute optimized:
\itemize{
\item Previous generation: (not recommended)

\strong{C1 node types:} \code{cache.c1.xlarge}
}
\item Memory optimized:
\itemize{
\item Current generation:

\strong{R3 node types:} \code{cache.r3.large}, \code{cache.r3.xlarge},
\code{cache.r3.2xlarge}, \code{cache.r3.4xlarge}, \code{cache.r3.8xlarge}

\strong{R4 node types;} \code{cache.r4.large}, \code{cache.r4.xlarge},
\code{cache.r4.2xlarge}, \code{cache.r4.4xlarge}, \code{cache.r4.8xlarge},
\code{cache.r4.16xlarge}
\item Previous generation: (not recommended)

\strong{M2 node types:} \code{cache.m2.xlarge}, \code{cache.m2.2xlarge},
\code{cache.m2.4xlarge}
}
}

\strong{Notes:}
\itemize{
\item All T2 instances are created in an Amazon Virtual Private Cloud
(Amazon VPC).
\item Redis (cluster mode disabled): Redis backup/restore is not supported
on T1 and T2 instances.
\item Redis (cluster mode enabled): Backup/restore is not supported on T1
instances.
\item Redis Append-only files (AOF) functionality is not supported for T1
or T2 instances.
}

For a complete listing of node types and specifications, see:
\itemize{
\item \href{http://aws.amazon.com/elasticache/details}{Amazon ElastiCache Product Features and Details}
\item \href{http://docs.aws.amazon.com/AmazonElastiCache/latest/mem-ug/ParameterGroups.Memcached.html#ParameterGroups.Memcached.NodeSpecific}{Cache Node Type-Specific Parameters for Memcached}
\item \href{http://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/ParameterGroups.Redis.html#ParameterGroups.Redis.NodeSpecific}{Cache Node Type-Specific Parameters for Redis}
}}

\item{Engine}{The name of the cache engine to be used for the clusters in this
replication group.}

\item{EngineVersion}{The version number of the cache engine to be used for the clusters in
this replication group. To view the supported cache engine versions, use
the \code{DescribeCacheEngineVersions} operation.

\strong{Important:} You can upgrade to a newer engine version (see \href{http://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/SelectEngine.html#VersionManagement}{Selecting a Cache Engine and Version})
in the \emph{ElastiCache User Guide}, but you cannot downgrade to an earlier
engine version. If you want to use an earlier engine version, you must
delete the existing cluster or replication group and create it anew with
the earlier engine version.}

\item{CacheParameterGroupName}{The name of the parameter group to associate with this replication
group. If this argument is omitted, the default cache parameter group
for the specified engine is used.

If you are running Redis version 3.2.4 or later, only one node group
(shard), and want to use a default parameter group, we recommend that
you specify the parameter group by name.
\itemize{
\item To create a Redis (cluster mode disabled) replication group, use
\code{CacheParameterGroupName=default.redis3.2}.
\item To create a Redis (cluster mode enabled) replication group, use
\code{CacheParameterGroupName=default.redis3.2.cluster.on}.
}}

\item{CacheSubnetGroupName}{The name of the cache subnet group to be used for the replication group.

If you're going to launch your cluster in an Amazon VPC, you need to
create a subnet group before you start creating a cluster. For more
information, see \href{http://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/SubnetGroups.html}{Subnets and Subnet Groups}.}

\item{CacheSecurityGroupNames}{A list of cache security group names to associate with this replication
group.}

\item{SecurityGroupIds}{One or more Amazon VPC security groups associated with this replication
group.

Use this parameter only when you are creating a replication group in an
Amazon Virtual Private Cloud (Amazon VPC).}

\item{Tags}{A list of cost allocation tags to be added to this resource. A tag is a
key-value pair.}

\item{SnapshotArns}{A list of Amazon Resource Names (ARN) that uniquely identify the Redis
RDB snapshot files stored in Amazon S3. The snapshot files are used to
populate the new replication group. The Amazon S3 object name in the ARN
cannot contain any commas. The new replication group will have the
number of node groups (console: shards) specified by the parameter
\emph{NumNodeGroups} or the number of node groups configured by
\emph{NodeGroupConfiguration} regardless of the number of ARNs specified
here.

Example of an Amazon S3 ARN: \code{arn:aws:s3:::my_bucket/snapshot1.rdb}}

\item{SnapshotName}{The name of a snapshot from which to restore data into the new
replication group. The snapshot status changes to \code{restoring} while the
new replication group is being created.}

\item{PreferredMaintenanceWindow}{Specifies the weekly time range during which maintenance on the cluster
is performed. It is specified as a range in the format
ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window
is a 60 minute period. Valid values for \code{ddd} are:

Specifies the weekly time range during which maintenance on the cluster
is performed. It is specified as a range in the format
ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window
is a 60 minute period.

Valid values for \code{ddd} are:
\itemize{
\item \code{sun}
\item \code{mon}
\item \code{tue}
\item \code{wed}
\item \code{thu}
\item \code{fri}
\item \code{sat}
}

Example: \code{sun:23:00-mon:01:30}}

\item{Port}{The port number on which each member of the replication group accepts
connections.}

\item{NotificationTopicArn}{The Amazon Resource Name (ARN) of the Amazon Simple Notification Service
(SNS) topic to which notifications are sent.

The Amazon SNS topic owner must be the same as the cluster owner.}

\item{AutoMinorVersionUpgrade}{This parameter is currently disabled.}

\item{SnapshotRetentionLimit}{The number of days for which ElastiCache retains automatic snapshots
before deleting them. For example, if you set \code{SnapshotRetentionLimit}
to 5, a snapshot that was taken today is retained for 5 days before
being deleted.

Default: 0 (i.e., automatic backups are disabled for this cluster).}

\item{SnapshotWindow}{The daily time range (in UTC) during which ElastiCache begins taking a
daily snapshot of your node group (shard).

Example: \code{05:00-09:00}

If you do not specify this parameter, ElastiCache automatically chooses
an appropriate time range.}

\item{AuthToken}{\strong{Reserved parameter.} The password used to access a password protected
server.

\code{AuthToken} can be specified only on replication groups where
\code{TransitEncryptionEnabled} is \code{true}.

For HIPAA compliance, you must specify \code{TransitEncryptionEnabled} as
\code{true}, an \code{AuthToken}, and a \code{CacheSubnetGroup}.

Password constraints:
\itemize{
\item Must be only printable ASCII characters.
\item Must be at least 16 characters and no more than 128 characters in
length.
\item Cannot contain any of the following characters: '/', '"', or
'@'.
}

For more information, see AUTH password at
http://redis.io/commands/AUTH.}

\item{TransitEncryptionEnabled}{A flag that enables in-transit encryption when set to \code{true}.

You cannot modify the value of \code{TransitEncryptionEnabled} after the
cluster is created. To enable in-transit encryption on a cluster you
must set \code{TransitEncryptionEnabled} to \code{true} when you create a cluster.

This parameter is valid only if the \code{Engine} parameter is \code{redis}, the
\code{EngineVersion} parameter is \code{3.2.6} or \code{4.x}, and the cluster is being
created in an Amazon VPC.

If you enable in-transit encryption, you must also specify a value for
\code{CacheSubnetGroup}.

\strong{Required:} Only available when creating a replication group in an
Amazon VPC using redis version \code{3.2.6} or \code{4.x}.

Default: \code{false}

For HIPAA compliance, you must specify \code{TransitEncryptionEnabled} as
\code{true}, an \code{AuthToken}, and a \code{CacheSubnetGroup}.}

\item{AtRestEncryptionEnabled}{A flag that enables encryption at rest when set to \code{true}.

You cannot modify the value of \code{AtRestEncryptionEnabled} after the
replication group is created. To enable encryption at rest on a
replication group you must set \code{AtRestEncryptionEnabled} to \code{true} when
you create the replication group.

\strong{Required:} Only available when creating a replication group in an
Amazon VPC using redis version \code{3.2.6} or \code{4.x}.

Default: \code{false}}
}
\description{
Creates a Redis (cluster mode disabled) or a Redis (cluster mode
enabled) replication group.
}
\details{
A Redis (cluster mode disabled) replication group is a collection of
clusters, where one of the clusters is a read/write primary and the
others are read-only replicas. Writes to the primary are asynchronously
propagated to the replicas.

A Redis (cluster mode enabled) replication group is a collection of 1 to
15 node groups (shards). Each node group (shard) has one read/write
primary node and up to 5 read-only replica nodes. Writes to the primary
are asynchronously propagated to the replicas. Redis (cluster mode
enabled) replication groups partition the data across node groups
(shards).

When a Redis (cluster mode disabled) replication group has been
successfully created, you can add one or more read replicas to it, up to
a total of 5 read replicas. You cannot alter a Redis (cluster mode
enabled) replication group after it has been created. However, if you
need to increase or decrease the number of node groups (console:
shards), you can avail yourself of ElastiCache for Redis' enhanced
backup and restore. For more information, see \href{http://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/backups-restoring.html}{Restoring From a Backup with Cluster Resizing}
in the \emph{ElastiCache User Guide}.

This operation is valid for Redis only.
}
\section{Request syntax}{
\preformatted{elasticache$create_replication_group(
  ReplicationGroupId = "string",
  ReplicationGroupDescription = "string",
  PrimaryClusterId = "string",
  AutomaticFailoverEnabled = TRUE|FALSE,
  NumCacheClusters = 123,
  PreferredCacheClusterAZs = list(
    "string"
  ),
  NumNodeGroups = 123,
  ReplicasPerNodeGroup = 123,
  NodeGroupConfiguration = list(
    list(
      NodeGroupId = "string",
      Slots = "string",
      ReplicaCount = 123,
      PrimaryAvailabilityZone = "string",
      ReplicaAvailabilityZones = list(
        "string"
      )
    )
  ),
  CacheNodeType = "string",
  Engine = "string",
  EngineVersion = "string",
  CacheParameterGroupName = "string",
  CacheSubnetGroupName = "string",
  CacheSecurityGroupNames = list(
    "string"
  ),
  SecurityGroupIds = list(
    "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  SnapshotArns = list(
    "string"
  ),
  SnapshotName = "string",
  PreferredMaintenanceWindow = "string",
  Port = 123,
  NotificationTopicArn = "string",
  AutoMinorVersionUpgrade = TRUE|FALSE,
  SnapshotRetentionLimit = 123,
  SnapshotWindow = "string",
  AuthToken = "string",
  TransitEncryptionEnabled = TRUE|FALSE,
  AtRestEncryptionEnabled = TRUE|FALSE
)
}
}

\keyword{internal}
