% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_create_snapshot}
\alias{elasticache_create_snapshot}
\title{Creates a copy of an entire cluster or replication group at a specific
moment in time}
\usage{
elasticache_create_snapshot(ReplicationGroupId, CacheClusterId,
  SnapshotName, KmsKeyId)
}
\arguments{
\item{ReplicationGroupId}{The identifier of an existing replication group. The snapshot is created
from this replication group.}

\item{CacheClusterId}{The identifier of an existing cluster. The snapshot is created from this
cluster.}

\item{SnapshotName}{[required] A name for the snapshot being created.}

\item{KmsKeyId}{The ID of the KMS key used to encrypt the snapshot.}
}
\description{
Creates a copy of an entire cluster or replication group at a specific
moment in time.
}
\details{
This operation is valid for Redis only.
}
\section{Request syntax}{
\preformatted{svc$create_snapshot(
  ReplicationGroupId = "string",
  CacheClusterId = "string",
  SnapshotName = "string",
  KmsKeyId = "string"
)
}
}

\keyword{internal}
