% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_snapshot}
\alias{elasticache_delete_snapshot}
\title{Deletes an existing snapshot}
\usage{
elasticache_delete_snapshot(SnapshotName)
}
\arguments{
\item{SnapshotName}{[required] The name of the snapshot to be deleted.}
}
\description{
Deletes an existing snapshot. When you receive a successful response
from this operation, ElastiCache immediately begins deleting the
snapshot; you cannot cancel or revert this operation.
}
\details{
This operation is valid for Redis only.
}
\section{Request syntax}{
\preformatted{elasticache$delete_snapshot(
  SnapshotName = "string"
)
}
}

\keyword{internal}
