% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_user_group}
\alias{elasticache_delete_user_group}
\title{For Redis engine version 6}
\usage{
elasticache_delete_user_group(UserGroupId)
}
\arguments{
\item{UserGroupId}{[required] The ID of the user group.}
}
\description{
For Redis engine version 6.x onwards: Deletes a ser group. The user
group must first be disassociated from the replcation group before it
can be deleted. For more information, see \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/Clusters.RBAC.html}{Using Role Based Access Control (RBAC)}.
}
\section{Request syntax}{
\preformatted{svc$delete_user_group(
  UserGroupId = "string"
)
}
}

\keyword{internal}
