% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_disassociate_global_replication_group}
\alias{elasticache_disassociate_global_replication_group}
\title{Remove a secondary cluster from the Global Datastore using the Global
Datastore name}
\usage{
elasticache_disassociate_global_replication_group(
  GlobalReplicationGroupId, ReplicationGroupId, ReplicationGroupRegion)
}
\arguments{
\item{GlobalReplicationGroupId}{[required] The name of the Global Datastore}

\item{ReplicationGroupId}{[required] The name of the secondary cluster you wish to remove from the Global
Datastore}

\item{ReplicationGroupRegion}{[required] The AWS region of secondary cluster you wish to remove from the Global
Datastore}
}
\description{
Remove a secondary cluster from the Global Datastore using the Global
Datastore name. The secondary cluster will no longer receive updates
from the primary cluster, but will remain as a standalone cluster in
that AWS region.
}
\section{Request syntax}{
\preformatted{svc$disassociate_global_replication_group(
  GlobalReplicationGroupId = "string",
  ReplicationGroupId = "string",
  ReplicationGroupRegion = "string"
)
}
}

\keyword{internal}
