% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_check_dns_availability}
\alias{elasticbeanstalk_check_dns_availability}
\title{Checks if the specified CNAME is available}
\usage{
elasticbeanstalk_check_dns_availability(CNAMEPrefix)
}
\arguments{
\item{CNAMEPrefix}{[required] The prefix used when this CNAME is reserved.}
}
\description{
Checks if the specified CNAME is available.
}
\section{Request syntax}{
\preformatted{elasticbeanstalk$check_dns_availability(
  CNAMEPrefix = "string"
)
}
}

\examples{
# The following operation checks the availability of the subdomain
# my-cname:
\donttest{elasticbeanstalk$check_dns_availability(
  CNAMEPrefix = "my-cname"
)}

}
\keyword{internal}
