% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_create_application}
\alias{elasticbeanstalk_create_application}
\title{Creates an application that has one configuration template named default
and no application versions}
\usage{
elasticbeanstalk_create_application(ApplicationName, Description,
  ResourceLifecycleConfig)
}
\arguments{
\item{ApplicationName}{[required] The name of the application.

Constraint: This name must be unique within your account. If the
specified name already exists, the action returns an
\code{InvalidParameterValue} error.}

\item{Description}{Describes the application.}

\item{ResourceLifecycleConfig}{Specify an application resource lifecycle configuration to prevent your
application from accumulating too many versions.}
}
\description{
Creates an application that has one configuration template named
\code{default} and no application versions.
}
\section{Request syntax}{
\preformatted{elasticbeanstalk$create_application(
  ApplicationName = "string",
  Description = "string",
  ResourceLifecycleConfig = list(
    ServiceRole = "string",
    VersionLifecycleConfig = list(
      MaxCountRule = list(
        Enabled = TRUE|FALSE,
        MaxCount = 123,
        DeleteSourceFromS3 = TRUE|FALSE
      ),
      MaxAgeRule = list(
        Enabled = TRUE|FALSE,
        MaxAgeInDays = 123,
        DeleteSourceFromS3 = TRUE|FALSE
      )
    )
  )
)
}
}

\examples{
# The following operation creates a new application named my-app:
\donttest{elasticbeanstalk$create_application(
  ApplicationName = "my-app",
  Description = "my application"
)}

}
\keyword{internal}
