% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_describe_account_attributes}
\alias{elasticbeanstalk_describe_account_attributes}
\title{Returns attributes related to AWS Elastic Beanstalk that are associated
with the calling AWS account}
\usage{
elasticbeanstalk_describe_account_attributes()
}
\description{
Returns attributes related to AWS Elastic Beanstalk that are associated
with the calling AWS account.
}
\details{
The result currently has one set of attributes---resource quotas.
}
\keyword{internal}
