% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_describe_application_versions}
\alias{elasticbeanstalk_describe_application_versions}
\title{Retrieve a list of application versions}
\usage{
elasticbeanstalk_describe_application_versions(ApplicationName,
  VersionLabels, MaxRecords, NextToken)
}
\arguments{
\item{ApplicationName}{Specify an application name to show only application versions for that
application.}

\item{VersionLabels}{Specify a version label to show a specific application version.}

\item{MaxRecords}{For a paginated request. Specify a maximum number of application
versions to include in each response.

If no \code{MaxRecords} is specified, all available application versions are
retrieved in a single response.}

\item{NextToken}{For a paginated request. Specify a token from a previous response page
to retrieve the next response page. All other parameter values must be
identical to the ones specified in the initial request.

If no \code{NextToken} is specified, the first page is retrieved.}
}
\description{
Retrieve a list of application versions.
}
\section{Request syntax}{
\preformatted{svc$describe_application_versions(
  ApplicationName = "string",
  VersionLabels = list(
    "string"
  ),
  MaxRecords = 123,
  NextToken = "string"
)
}
}

\examples{
# The following operation retrieves information about an application
# version labeled v2:
\donttest{svc$describe_application_versions(
  ApplicationName = "my-app",
  VersionLabels = list(
    "v2"
  )
)}

}
\keyword{internal}
