% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_describe_environment_managed_actions}
\alias{elasticbeanstalk_describe_environment_managed_actions}
\title{Lists an environment's upcoming and in-progress managed actions}
\usage{
elasticbeanstalk_describe_environment_managed_actions(EnvironmentName,
  EnvironmentId, Status)
}
\arguments{
\item{EnvironmentName}{The name of the target environment.}

\item{EnvironmentId}{The environment ID of the target environment.}

\item{Status}{To show only actions with a particular status, specify a status.}
}
\description{
Lists an environment's upcoming and in-progress managed actions.
}
\section{Request syntax}{
\preformatted{svc$describe_environment_managed_actions(
  EnvironmentName = "string",
  EnvironmentId = "string",
  Status = "Scheduled"|"Pending"|"Running"|"Unknown"
)
}
}

\keyword{internal}
