% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_describe_platform_version}
\alias{elasticbeanstalk_describe_platform_version}
\title{Describes the version of the platform}
\usage{
elasticbeanstalk_describe_platform_version(PlatformArn)
}
\arguments{
\item{PlatformArn}{The ARN of the version of the platform.}
}
\description{
Describes the version of the platform.
}
\section{Request syntax}{
\preformatted{svc$describe_platform_version(
  PlatformArn = "string"
)
}
}

\keyword{internal}
