% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_list_available_solution_stacks}
\alias{elasticbeanstalk_list_available_solution_stacks}
\title{Returns a list of the available solution stack names, with the public
version first and then in reverse chronological order}
\usage{
elasticbeanstalk_list_available_solution_stacks()
}
\description{
Returns a list of the available solution stack names, with the public
version first and then in reverse chronological order.
}
\examples{
# The following operation lists solution stacks for all currently
# available platform configurations and any that you have used in the
# past:
\donttest{svc$list_available_solution_stacks()}

}
\keyword{internal}
