% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_service.R
\name{elasticinference}
\alias{elasticinference}
\title{Amazon Elastic  Inference}
\usage{
elasticinference(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Elastic Inference public APIs.
}
\section{Service syntax}{
\preformatted{svc <- elasticinference(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Examples}{
\preformatted{svc$list_tags_for_resource(
  Foo = 123
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=elasticinference_list_tags_for_resource]{list_tags_for_resource} \tab Returns all tags of an Elastic Inference Accelerator \cr
\link[=elasticinference_tag_resource]{tag_resource} \tab Adds the specified tag(s) to an Elastic Inference Accelerator \cr
\link[=elasticinference_untag_resource]{untag_resource} \tab Removes the specified tag(s) from an Elastic Inference Accelerator
}
}

