% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_operations.R
\name{elasticinference_describe_accelerator_offerings}
\alias{elasticinference_describe_accelerator_offerings}
\title{Describes the locations in which a given accelerator type or set of
types is present in a given region}
\usage{
elasticinference_describe_accelerator_offerings(locationType,
  acceleratorTypes)
}
\arguments{
\item{locationType}{[required] The location type that you want to describe accelerator type offerings
for. It can assume the following values: region: will return the
accelerator type offering at the regional level. availability-zone: will
return the accelerator type offering at the availability zone level.
availability-zone-id: will return the accelerator type offering at the
availability zone level returning the availability zone id.}

\item{acceleratorTypes}{The list of accelerator types to describe.}
}
\description{
Describes the locations in which a given accelerator type or set of
types is present in a given region.
}
\section{Request syntax}{
\preformatted{svc$describe_accelerator_offerings(
  locationType = "region"|"availability-zone"|"availability-zone-id",
  acceleratorTypes = list(
    "string"
  )
)
}
}

\keyword{internal}
