% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_operations.R
\name{elasticinference_describe_accelerator_types}
\alias{elasticinference_describe_accelerator_types}
\title{Describes the accelerator types available in a given region, as well as
their characteristics, such as memory and throughput}
\usage{
elasticinference_describe_accelerator_types()
}
\description{
Describes the accelerator types available in a given region, as well as
their characteristics, such as memory and throughput.
}
\section{Request syntax}{
\preformatted{svc$describe_accelerator_types()
}
}

\keyword{internal}
