% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_operations.R
\name{elasticinference_describe_accelerators}
\alias{elasticinference_describe_accelerators}
\title{Describes information over a provided set of accelerators belonging to
an account}
\usage{
elasticinference_describe_accelerators(acceleratorIds, filters,
  maxResults, nextToken)
}
\arguments{
\item{acceleratorIds}{The IDs of the accelerators to describe.}

\item{filters}{One or more filters. Filter names and values are case-sensitive. Valid
filter names are: accelerator-types: can provide a list of accelerator
type names to filter for. instance-id: can provide a list of EC2
instance ids to filter for.}

\item{maxResults}{The total number of items to return in the command\'s output. If the
total number of items available is more than the value specified, a
NextToken is provided in the command\'s output. To resume pagination,
provide the NextToken value in the starting-token argument of a
subsequent command. Do not use the NextToken response element directly
outside of the AWS CLI.}

\item{nextToken}{A token to specify where to start paginating. This is the NextToken from
a previously truncated response.}
}
\description{
Describes information over a provided set of accelerators belonging to
an account.
}
\section{Request syntax}{
\preformatted{svc$describe_accelerators(
  acceleratorIds = list(
    "string"
  ),
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
