% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_accept_inbound_cross_cluster_search_connection}
\alias{elasticsearchservice_accept_inbound_cross_cluster_search_connection}
\title{Allows the destination domain owner to accept an inbound cross-cluster
search connection request}
\usage{
elasticsearchservice_accept_inbound_cross_cluster_search_connection(
  CrossClusterSearchConnectionId)
}
\arguments{
\item{CrossClusterSearchConnectionId}{[required] The id of the inbound connection that you want to accept.}
}
\description{
Allows the destination domain owner to accept an inbound cross-cluster
search connection request.
}
\section{Request syntax}{
\preformatted{svc$accept_inbound_cross_cluster_search_connection(
  CrossClusterSearchConnectionId = "string"
)
}
}

\keyword{internal}
