% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_create_outbound_cross_cluster_search_connection}
\alias{elasticsearchservice_create_outbound_cross_cluster_search_connection}
\title{Creates a new cross-cluster search connection from a source domain to a
destination domain}
\usage{
elasticsearchservice_create_outbound_cross_cluster_search_connection(
  SourceDomainInfo, DestinationDomainInfo, ConnectionAlias)
}
\arguments{
\item{SourceDomainInfo}{[required] Specifies the \verb{<a>DomainInformation</a>} for the source Elasticsearch
domain.}

\item{DestinationDomainInfo}{[required] Specifies the \verb{<a>DomainInformation</a>} for the destination
Elasticsearch domain.}

\item{ConnectionAlias}{[required] Specifies the connection alias that will be used by the customer for
this connection.}
}
\description{
Creates a new cross-cluster search connection from a source domain to a
destination domain.
}
\section{Request syntax}{
\preformatted{svc$create_outbound_cross_cluster_search_connection(
  SourceDomainInfo = list(
    OwnerId = "string",
    DomainName = "string",
    Region = "string"
  ),
  DestinationDomainInfo = list(
    OwnerId = "string",
    DomainName = "string",
    Region = "string"
  ),
  ConnectionAlias = "string"
)
}
}

\keyword{internal}
