% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_delete_package}
\alias{elasticsearchservice_delete_package}
\title{Delete the package}
\usage{
elasticsearchservice_delete_package(PackageID)
}
\arguments{
\item{PackageID}{[required] Internal ID of the package that you want to delete. Use
\code{DescribePackages} to find this value.}
}
\description{
Delete the package.
}
\section{Request syntax}{
\preformatted{svc$delete_package(
  PackageID = "string"
)
}
}

\keyword{internal}
