% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_packages}
\alias{elasticsearchservice_describe_packages}
\title{Describes all packages available to Amazon ES}
\usage{
elasticsearchservice_describe_packages(Filters, MaxResults, NextToken)
}
\arguments{
\item{Filters}{Only returns packages that match the \code{DescribePackagesFilterList}
values.}

\item{MaxResults}{Limits results to a maximum number of packages.}

\item{NextToken}{Used for pagination. Only necessary if a previous API call includes a
non-null NextToken value. If provided, returns results for the next
page.}
}
\description{
Describes all packages available to Amazon ES. Includes options for
filtering, limiting the number of results, and pagination.
}
\section{Request syntax}{
\preformatted{svc$describe_packages(
  Filters = list(
    list(
      Name = "PackageID"|"PackageName"|"PackageStatus",
      Value = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
