% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_reserved_elasticsearch_instances}
\alias{elasticsearchservice_describe_reserved_elasticsearch_instances}
\title{Returns information about reserved Elasticsearch instances for this
account}
\usage{
elasticsearchservice_describe_reserved_elasticsearch_instances(
  ReservedElasticsearchInstanceId, MaxResults, NextToken)
}
\arguments{
\item{ReservedElasticsearchInstanceId}{The reserved instance identifier filter value. Use this parameter to
show only the reservation that matches the specified reserved
Elasticsearch instance ID.}

\item{MaxResults}{Set this value to limit the number of results returned. If not
specified, defaults to 100.}

\item{NextToken}{NextToken should be sent in case if earlier API call produced result
containing NextToken. It is used for pagination.}
}
\description{
Returns information about reserved Elasticsearch instances for this
account.
}
\section{Request syntax}{
\preformatted{svc$describe_reserved_elasticsearch_instances(
  ReservedElasticsearchInstanceId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
