% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_get_upgrade_history}
\alias{elasticsearchservice_get_upgrade_history}
\title{Retrieves the complete history of the last 10 upgrades that were
performed on the domain}
\usage{
elasticsearchservice_get_upgrade_history(DomainName, MaxResults,
  NextToken)
}
\arguments{
\item{DomainName}{[required]}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Retrieves the complete history of the last 10 upgrades that were
performed on the domain.
}
\section{Request syntax}{
\preformatted{elasticsearchservice$get_upgrade_history(
  DomainName = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
