% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_purchase_reserved_elasticsearch_instance_offering}
\alias{elasticsearchservice_purchase_reserved_elasticsearch_instance_offering}
\title{Allows you to purchase reserved Elasticsearch instances}
\usage{
elasticsearchservice_purchase_reserved_elasticsearch_instance_offering(
  ReservedElasticsearchInstanceOfferingId, ReservationName, InstanceCount)
}
\arguments{
\item{ReservedElasticsearchInstanceOfferingId}{[required] The ID of the reserved Elasticsearch instance offering to purchase.}

\item{ReservationName}{[required] A customer-specified identifier to track this reservation.}

\item{InstanceCount}{The number of Elasticsearch instances to reserve.}
}
\description{
Allows you to purchase reserved Elasticsearch instances.
}
\section{Request syntax}{
\preformatted{elasticsearchservice$purchase_reserved_elasticsearch_instance_offering(
  ReservedElasticsearchInstanceOfferingId = "string",
  ReservationName = "string",
  InstanceCount = 123
)
}
}

\keyword{internal}
