% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_start_elasticsearch_service_software_update}
\alias{elasticsearchservice_start_elasticsearch_service_software_update}
\title{Schedules a service software update for an Amazon ES domain}
\usage{
elasticsearchservice_start_elasticsearch_service_software_update(
  DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to update to the latest service
software.}
}
\description{
Schedules a service software update for an Amazon ES domain.
}
\section{Request syntax}{
\preformatted{elasticsearchservice$start_elasticsearch_service_software_update(
  DomainName = "string"
)
}
}

\keyword{internal}
