% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_read_job}
\alias{elastictranscoder_read_job}
\title{The ReadJob operation returns detailed information about a job}
\usage{
elastictranscoder_read_job(Id)
}
\arguments{
\item{Id}{[required] The identifier of the job for which you want to get detailed
information.}
}
\description{
The ReadJob operation returns detailed information about a job.
}
\section{Request syntax}{
\preformatted{elastictranscoder$read_job(
  Id = "string"
)
}
}

\keyword{internal}
