% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_delete_rule}
\alias{elbv2_delete_rule}
\title{Deletes the specified rule}
\usage{
elbv2_delete_rule(RuleArn)
}
\arguments{
\item{RuleArn}{[required] The Amazon Resource Name (ARN) of the rule.}
}
\description{
Deletes the specified rule.
}
\details{
You can't delete the default rule.
}
\section{Request syntax}{
\preformatted{svc$delete_rule(
  RuleArn = "string"
)
}
}

\examples{
\dontrun{
# This example deletes the specified rule.
svc$delete_rule(
  RuleArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:listener-rule/app/my-load-b..."
)
}

}
\keyword{internal}
