% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_listener_certificates}
\alias{elbv2_describe_listener_certificates}
\title{Describes the certificates for the specified HTTPS listener}
\usage{
elbv2_describe_listener_certificates(ListenerArn, Marker, PageSize)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Names (ARN) of the listener.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes the certificates for the specified HTTPS listener.
}
\section{Request syntax}{
\preformatted{elbv2$describe_listener_certificates(
  ListenerArn = "string",
  Marker = "string",
  PageSize = 123
)
}
}

\keyword{internal}
