% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_get_managed_scaling_policy}
\alias{emr_get_managed_scaling_policy}
\title{Fetches the attached managed scaling policy for an Amazon EMR cluster}
\usage{
emr_get_managed_scaling_policy(ClusterId)
}
\arguments{
\item{ClusterId}{[required] Specifies the ID of the cluster for which the managed scaling policy
will be fetched.}
}
\description{
Fetches the attached managed scaling policy for an Amazon EMR cluster.
}
\section{Request syntax}{
\preformatted{svc$get_managed_scaling_policy(
  ClusterId = "string"
)
}
}

\keyword{internal}
