% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_list_studios}
\alias{emr_list_studios}
\title{The Amazon EMR Studio APIs are in preview release for Amazon EMR and are
subject to change}
\usage{
emr_list_studios(Marker)
}
\arguments{
\item{Marker}{The pagination token that indicates the set of results to retrieve.}
}
\description{
The Amazon EMR Studio APIs are in preview release for Amazon EMR and are
subject to change.

Returns a list of all Amazon EMR Studios associated with the AWS
account. The list includes details such as ID, Studio Access URL, and
creation time for each Studio.
}
\section{Request syntax}{
\preformatted{svc$list_studios(
  Marker = "string"
)
}
}

\keyword{internal}
