% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_create_virtual_cluster}
\alias{emrcontainers_create_virtual_cluster}
\title{Creates a virtual cluster}
\usage{
emrcontainers_create_virtual_cluster(name, containerProvider,
  clientToken, tags)
}
\arguments{
\item{name}{[required] The specified name of the virtual cluster.}

\item{containerProvider}{[required] The container provider of the virtual cluster.}

\item{clientToken}{[required] The client token of the virtual cluster.}

\item{tags}{The tags assigned to the virtual cluster.}
}
\description{
Creates a virtual cluster. Virtual cluster is a managed entity on Amazon
EMR on EKS. You can create, describe, list and delete virtual clusters.
They do not consume any additional resource in your system. A single
virtual cluster maps to a single Kubernetes namespace. Given this
relationship, you can model virtual clusters the same way you model
Kubernetes namespaces to meet your requirements.
}
\section{Request syntax}{
\preformatted{svc$create_virtual_cluster(
  name = "string",
  containerProvider = list(
    type = "EKS",
    id = "string",
    info = list(
      eksInfo = list(
        namespace = "string"
      )
    )
  ),
  clientToken = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
