% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_untag_resource}
\alias{emrcontainers_untag_resource}
\title{Removes tags from resources}
\usage{
emrcontainers_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of resources.}

\item{tagKeys}{[required] The tag keys of the resources.}
}
\description{
Removes tags from resources.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
