% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_deactivate_event_source}
\alias{eventbridge_deactivate_event_source}
\title{You can use this operation to temporarily stop receiving events from the
specified partner event source}
\usage{
eventbridge_deactivate_event_source(Name)
}
\arguments{
\item{Name}{[required] The name of the partner event source to deactivate.}
}
\description{
You can use this operation to temporarily stop receiving events from the
specified partner event source. The matching event bus is not deleted.

When you deactivate a partner event source, the source goes into PENDING
state. If it remains in PENDING state for more than two weeks, it is
deleted.

To activate a deactivated partner event source, use ActivateEventSource.
}
\section{Request syntax}{
\preformatted{svc$deactivate_event_source(
  Name = "string"
)
}
}

\keyword{internal}
