% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_delete_partner_event_source}
\alias{eventbridge_delete_partner_event_source}
\title{This operation is used by SaaS partners to delete a partner event source}
\usage{
eventbridge_delete_partner_event_source(Name, Account)
}
\arguments{
\item{Name}{[required] The name of the event source to delete.}

\item{Account}{[required] The AWS account ID of the AWS customer that the event source was created
for.}
}
\description{
This operation is used by SaaS partners to delete a partner event
source. This operation is not used by AWS customers.
}
\details{
When you delete an event source, the status of the corresponding partner
event bus in the AWS customer account becomes DELETED.
}
\section{Request syntax}{
\preformatted{svc$delete_partner_event_source(
  Name = "string",
  Account = "string"
)
}
}

\keyword{internal}
