% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_tags_for_resource}
\alias{eventbridge_list_tags_for_resource}
\title{Displays the tags associated with an EventBridge resource}
\usage{
eventbridge_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the rule for which you want to view tags.}
}
\description{
Displays the tags associated with an EventBridge resource. In
EventBridge, rules can be tagged.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}
}

\keyword{internal}
