% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_put_events}
\alias{eventbridge_put_events}
\title{Sends custom events to EventBridge so that they can be matched to rules}
\usage{
eventbridge_put_events(Entries)
}
\arguments{
\item{Entries}{[required] The entry that defines an event in your system. You can specify several
parameters for the entry such as the source and type of the event,
resources associated with the event, and so on.}
}
\description{
Sends custom events to EventBridge so that they can be matched to rules.
These events can be from your custom applications and services.
}
\section{Request syntax}{
\preformatted{svc$put_events(
  Entries = list(
    list(
      Time = as.POSIXct(
        "2015-01-01"
      ),
      Source = "string",
      Resources = list(
        "string"
      ),
      DetailType = "string",
      Detail = "string",
      EventBusName = "string"
    )
  )
)
}
}

\keyword{internal}
