% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{extract_features}
\alias{extract_features}
\title{Extract features for paw time series}
\usage{
extract_features(x, y = NULL, parameters = default_parameters())
}
\arguments{
\item{x}{time series of horizontal paw movement. Alternatively,
a two column matrix of x and y time series, respectively.}

\item{y}{time series of vertical paw movement, or NULL if x is
a two column matrix.}

\item{parameters}{contains information about frames per second,
filtering, windowing, and thresholds, for paw features
(see \code{\link{default_parameters}}, or use
\code{\link{set_parameters}} to modify the defaults).}
}
\value{
pre-peak and post-peak paw features
}
\description{
Extract features for paw time series
}
