% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_delete_apps_list}
\alias{fms_delete_apps_list}
\title{Permanently deletes an AWS Firewall Manager applications list}
\usage{
fms_delete_apps_list(ListId)
}
\arguments{
\item{ListId}{[required] The ID of the applications list that you want to delete. You can
retrieve this ID from \code{PutAppsList}, \code{ListAppsLists}, and \code{GetAppsList}.}
}
\description{
Permanently deletes an AWS Firewall Manager applications list.
}
\section{Request syntax}{
\preformatted{svc$delete_apps_list(
  ListId = "string"
)
}
}

\keyword{internal}
