% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_policy}
\alias{fms_get_policy}
\title{Returns information about the specified AWS Firewall Manager policy}
\usage{
fms_get_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] The ID of the AWS Firewall Manager policy that you want the details for.}
}
\description{
Returns information about the specified AWS Firewall Manager policy.
}
\section{Request syntax}{
\preformatted{fms$get_policy(
  PolicyId = "string"
)
}
}

\keyword{internal}
