% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_list_tags_for_resource}
\alias{fms_list_tags_for_resource}
\title{Retrieves the list of tags for the specified AWS resource}
\usage{
fms_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to return tags for. The
Firewall Manager policy is the only AWS resource that supports tagging,
so this ARN is a policy ARN..}
}
\description{
Retrieves the list of tags for the specified AWS resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
