% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_put_notification_channel}
\alias{fms_put_notification_channel}
\title{Designates the IAM role and Amazon Simple Notification Service (SNS)
topic that AWS Firewall Manager uses to record SNS logs}
\usage{
fms_put_notification_channel(SnsTopicArn, SnsRoleName)
}
\arguments{
\item{SnsTopicArn}{[required] The Amazon Resource Name (ARN) of the SNS topic that collects
notifications from AWS Firewall Manager.}

\item{SnsRoleName}{[required] The Amazon Resource Name (ARN) of the IAM role that allows Amazon SNS to
record AWS Firewall Manager activity.}
}
\description{
Designates the IAM role and Amazon Simple Notification Service (SNS)
topic that AWS Firewall Manager uses to record SNS logs.
}
\section{Request syntax}{
\preformatted{svc$put_notification_channel(
  SnsTopicArn = "string",
  SnsRoleName = "string"
)
}
}

\keyword{internal}
