% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_put_protocols_list}
\alias{fms_put_protocols_list}
\title{Creates an AWS Firewall Manager protocols list}
\usage{
fms_put_protocols_list(ProtocolsList, TagList)
}
\arguments{
\item{ProtocolsList}{[required] The details of the AWS Firewall Manager protocols list to be created.}

\item{TagList}{The tags associated with the resource.}
}
\description{
Creates an AWS Firewall Manager protocols list.
}
\section{Request syntax}{
\preformatted{svc$put_protocols_list(
  ProtocolsList = list(
    ListId = "string",
    ListName = "string",
    ListUpdateToken = "string",
    CreateTime = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdateTime = as.POSIXct(
      "2015-01-01"
    ),
    ProtocolsList = list(
      "string"
    ),
    PreviousProtocolsList = list(
      list(
        "string"
      )
    )
  ),
  TagList = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
