% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_untag_resource}
\alias{fms_untag_resource}
\title{Removes one or more tags from an AWS resource}
\usage{
fms_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource. The Firewall Manager
policy is the only AWS resource that supports tagging, so this ARN is a
policy ARN.}

\item{TagKeys}{[required] The keys of the tags to remove from the resource.}
}
\description{
Removes one or more tags from an AWS resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
