% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastqueryservice_operations.R
\name{forecastqueryservice_query_forecast}
\alias{forecastqueryservice_query_forecast}
\title{Retrieves a forecast filtered by the supplied criteria}
\usage{
forecastqueryservice_query_forecast(ForecastArn, StartDate, EndDate,
  Filters, NextToken)
}
\arguments{
\item{ForecastArn}{[required] The Amazon Resource Name (ARN) of the forecast to query.}

\item{StartDate}{The start date for the forecast. Specify the date using this format:
yyyy-MM-dd\'T\'HH:mm:ss\'Z\' (ISO 8601 format) For example,
\"1970-01-01T00:00:00Z.\"}

\item{EndDate}{The end date for the forecast. Specify the date using this format:
yyyy-MM-dd\'T\'HH:mm:ss\'Z\' (ISO 8601 format). For example,
\"1970-01-01T00:00:00Z.\"}

\item{Filters}{[required] The filtering criteria to apply when retrieving the forecast. For
example:
\itemize{
\item To get a forecast for a specific item specify the following:

\verb{\\\{"item_id" : "client_1"\\\}}
}<!-- -->

\itemize{
\item To get a forecast for a specific item sold in a specific location,
specify the following:

\verb{\\\{"item_id" : "client_1", "location" : "ny"\\\}}
}<!-- -->

\itemize{
\item To get a forecast for all blue items sold in a specific location,
specify the following:

\verb{\\\{ "location" : "ny", "color":"blue"\\\}}
}

To get the full forecast, use the operation.}

\item{NextToken}{If the result of the previous request was truncated, the response
includes a \code{NextToken}. To retrieve the next set of results, use the
token in the next request. Tokens expire after 24 hours.}
}
\description{
Retrieves a forecast filtered by the supplied criteria.
}
\details{
The criteria is a key-value pair. The key is either \code{item_id} (or the
equivalent non-timestamp, non-target field) from the
\code{TARGET_TIME_SERIES} dataset, or one of the forecast dimensions
specified as part of the \code{FeaturizationConfig} object.

By default, the complete date range of the filtered forecast is
returned. Optionally, you can request a specific date range within the
forecast.

The forecasts generated by Amazon Forecast are in the same timezone as
the dataset that was used to create the predictor.
}
\section{Request syntax}{
\preformatted{svc$query_forecast(
  ForecastArn = "string",
  StartDate = "string",
  EndDate = "string",
  Filters = list(
    "string"
  ),
  NextToken = "string"
)
}
}

\keyword{internal}
