% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_create_dataset}
\alias{forecastservice_create_dataset}
\title{Creates an Amazon Forecast dataset}
\usage{
forecastservice_create_dataset(DatasetName, Domain, DatasetType,
  DataFrequency, Schema, EncryptionConfig)
}
\arguments{
\item{DatasetName}{[required] A name for the dataset.}

\item{Domain}{[required] The domain associated with the dataset. When you add a dataset to a
dataset group, this value and the value specified for the \code{Domain}
parameter of the CreateDatasetGroup operation must match.

The \code{Domain} and \code{DatasetType} that you choose determine the fields that
must be present in the training data that you import to the dataset. For
example, if you choose the \code{RETAIL} domain and \code{TARGET_TIME_SERIES} as
the \code{DatasetType}, Amazon Forecast requires \code{item_id}, \code{timestamp}, and
\code{demand} fields to be present in your data. For more information, see
howitworks-datasets-groups.}

\item{DatasetType}{[required] The dataset type. Valid values depend on the chosen \code{Domain}.}

\item{DataFrequency}{The frequency of data collection. This parameter is required for
RELATED\\_TIME\\_SERIES datasets.

Valid intervals are Y (Year), M (Month), W (Week), D (Day), H (Hour),
30min (30 minutes), 15min (15 minutes), 10min (10 minutes), 5min (5
minutes), and 1min (1 minute). For example, \"D\" indicates every day
and \"15min\" indicates every 15 minutes.}

\item{Schema}{[required] The schema for the dataset. The schema attributes and their order must
match the fields in your data. The dataset \code{Domain} and \code{DatasetType}
that you choose determine the minimum required fields in your training
data. For information about the required fields for a specific dataset
domain and type, see howitworks-domains-ds-types.}

\item{EncryptionConfig}{An AWS Key Management Service (KMS) key and the AWS Identity and Access
Management (IAM) role that Amazon Forecast can assume to access the key.}
}
\description{
Creates an Amazon Forecast dataset. The information about the dataset
that you provide helps Forecast understand how to consume the data for
model training. This includes the following:
}
\details{
\itemize{
\item \emph{\code{DataFrequency}} - How frequently your historical time-series data
is collected.
\item \emph{\code{Domain}} and \emph{\code{DatasetType}} - Each dataset has an associated
dataset domain and a type within the domain. Amazon Forecast
provides a list of predefined domains and types within each domain.
For each unique dataset domain and type within the domain, Amazon
Forecast requires your data to include a minimum set of predefined
fields.
\item \emph{\code{Schema}} - A schema specifies the fields in the dataset, including
the field name and data type.
}

After creating a dataset, you import your training data into it and add
the dataset to a dataset group. You use the dataset group to create a
predictor. For more information, see howitworks-datasets-groups.

To get a list of all your datasets, use the ListDatasets operation.

For example Forecast datasets, see the Amazon Forecast Sample GitHub
repository.

The \code{Status} of a dataset must be \code{ACTIVE} before you can import
training data. Use the DescribeDataset operation to get the status.
}
\section{Request syntax}{
\preformatted{svc$create_dataset(
  DatasetName = "string",
  Domain = "RETAIL"|"CUSTOM"|"INVENTORY_PLANNING"|"EC2_CAPACITY"|"WORK_FORCE"|"WEB_TRAFFIC"|"METRICS",
  DatasetType = "TARGET_TIME_SERIES"|"RELATED_TIME_SERIES"|"ITEM_METADATA",
  DataFrequency = "string",
  Schema = list(
    Attributes = list(
      list(
        AttributeName = "string",
        AttributeType = "string"|"integer"|"float"|"timestamp"
      )
    )
  ),
  EncryptionConfig = list(
    RoleArn = "string",
    KMSKeyArn = "string"
  )
)
}
}

\keyword{internal}
