% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_create_dataset_group}
\alias{forecastservice_create_dataset_group}
\title{Creates a dataset group, which holds a collection of related datasets}
\usage{
forecastservice_create_dataset_group(DatasetGroupName, Domain,
  DatasetArns)
}
\arguments{
\item{DatasetGroupName}{[required] A name for the dataset group.}

\item{Domain}{[required] The domain associated with the dataset group. When you add a dataset to
a dataset group, this value and the value specified for the \code{Domain}
parameter of the CreateDataset operation must match.

The \code{Domain} and \code{DatasetType} that you choose determine the fields that
must be present in training data that you import to a dataset. For
example, if you choose the \code{RETAIL} domain and \code{TARGET_TIME_SERIES} as
the \code{DatasetType}, Amazon Forecast requires that \code{item_id}, \code{timestamp},
and \code{demand} fields are present in your data. For more information, see
howitworks-datasets-groups.}

\item{DatasetArns}{An array of Amazon Resource Names (ARNs) of the datasets that you want
to include in the dataset group.}
}
\description{
Creates a dataset group, which holds a collection of related datasets.
You can add datasets to the dataset group when you create the dataset
group, or later by using the UpdateDatasetGroup operation.
}
\details{
After creating a dataset group and adding datasets, you use the dataset
group when you create a predictor. For more information, see
howitworks-datasets-groups.

To get a list of all your datasets groups, use the ListDatasetGroups
operation.

The \code{Status} of a dataset group must be \code{ACTIVE} before you can create
use the dataset group to create a predictor. To get the status, use the
DescribeDatasetGroup operation.
}
\section{Request syntax}{
\preformatted{svc$create_dataset_group(
  DatasetGroupName = "string",
  Domain = "RETAIL"|"CUSTOM"|"INVENTORY_PLANNING"|"EC2_CAPACITY"|"WORK_FORCE"|"WEB_TRAFFIC"|"METRICS",
  DatasetArns = list(
    "string"
  )
)
}
}

\keyword{internal}
