% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_create_dataset_import_job}
\alias{forecastservice_create_dataset_import_job}
\title{Imports your training data to an Amazon Forecast dataset}
\usage{
forecastservice_create_dataset_import_job(DatasetImportJobName,
  DatasetArn, DataSource, TimestampFormat)
}
\arguments{
\item{DatasetImportJobName}{[required] The name for the dataset import job. We recommend including the current
timestamp in the name, for example, \verb{20190721DatasetImport}. This can
help you avoid getting a \code{ResourceAlreadyExistsException} exception.}

\item{DatasetArn}{[required] The Amazon Resource Name (ARN) of the Amazon Forecast dataset that you
want to import data to.}

\item{DataSource}{[required] The location of the training data to import and an AWS Identity and
Access Management (IAM) role that Amazon Forecast can assume to access
the data. The training data must be stored in an Amazon S3 bucket.

If encryption is used, \code{DataSource} must include an AWS Key Management
Service (KMS) key and the IAM role must allow Amazon Forecast permission
to access the key. The KMS key and IAM role must match those specified
in the \code{EncryptionConfig} parameter of the CreateDataset operation.}

\item{TimestampFormat}{The format of timestamps in the dataset. The format that you specify
depends on the \code{DataFrequency} specified when the dataset was created.
The following formats are supported
\itemize{
\item \"yyyy-MM-dd\"

For the following data frequencies: Y, M, W, and D
\item \"yyyy-MM-dd HH:mm:ss\"

For the following data frequencies: H, 30min, 15min, and 1min; and
optionally, for: Y, M, W, and D
}

If the format isn\'t specified, Amazon Forecast expects the format to be
\"yyyy-MM-dd HH:mm:ss\".}
}
\description{
Imports your training data to an Amazon Forecast dataset. You provide
the location of your training data in an Amazon Simple Storage Service
(Amazon S3) bucket and the Amazon Resource Name (ARN) of the dataset
that you want to import the data to.
}
\details{
You must specify a DataSource object that includes an AWS Identity and
Access Management (IAM) role that Amazon Forecast can assume to access
the data. For more information, see aws-forecast-iam-roles.

The training data must be in CSV format. The delimiter must be a comma
(,).

You can specify the path to a specific CSV file, the S3 bucket, or to a
folder in the S3 bucket. For the latter two cases, Amazon Forecast
imports all files up to the limit of 10,000 files.

To get a list of all your dataset import jobs, filtered by specified
criteria, use the ListDatasetImportJobs operation.
}
\section{Request syntax}{
\preformatted{svc$create_dataset_import_job(
  DatasetImportJobName = "string",
  DatasetArn = "string",
  DataSource = list(
    S3Config = list(
      Path = "string",
      RoleArn = "string",
      KMSKeyArn = "string"
    )
  ),
  TimestampFormat = "string"
)
}
}

\keyword{internal}
