% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_create_forecast}
\alias{forecastservice_create_forecast}
\title{Creates a forecast for each item in the TARGET_TIME_SERIES dataset that
was used to train the predictor}
\usage{
forecastservice_create_forecast(ForecastName, PredictorArn,
  ForecastTypes)
}
\arguments{
\item{ForecastName}{[required] A name for the forecast.}

\item{PredictorArn}{[required] The Amazon Resource Name (ARN) of the predictor to use to generate the
forecast.}

\item{ForecastTypes}{The quantiles at which probabilistic forecasts are generated. You can
specify up to 5 quantiles per forecast. Accepted values include
\verb{0.01 to 0.99} (increments of .01 only) and \code{mean}. The mean forecast is
different from the median (0.50) when the distribution is not symmetric
(e.g. Beta, Negative Binomial). The default value is
\verb{\\\["0.1", "0.5", "0.9"\\\]}.}
}
\description{
Creates a forecast for each item in the \code{TARGET_TIME_SERIES} dataset
that was used to train the predictor. This is known as inference. To
retrieve the forecast for a single item at low latency, use the
operation. To export the complete forecast into your Amazon Simple
Storage Service (Amazon S3) bucket, use the CreateForecastExportJob
operation.
}
\details{
The range of the forecast is determined by the \code{ForecastHorizon} value,
which you specify in the CreatePredictor request, multiplied by the
\code{DataFrequency} value, which you specify in the CreateDataset request.
When you query a forecast, you can request a specific date range within
the forecast.

To get a list of all your forecasts, use the ListForecasts operation.

The forecasts generated by Amazon Forecast are in the same time zone as
the dataset that was used to create the predictor.

For more information, see howitworks-forecast.

The \code{Status} of the forecast must be \code{ACTIVE} before you can query or
export the forecast. Use the DescribeForecast operation to get the
status.
}
\section{Request syntax}{
\preformatted{svc$create_forecast(
  ForecastName = "string",
  PredictorArn = "string",
  ForecastTypes = list(
    "string"
  )
)
}
}

\keyword{internal}
