% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_dataset}
\alias{forecastservice_delete_dataset}
\title{Deletes an Amazon Forecast dataset that was created using the
CreateDataset operation}
\usage{
forecastservice_delete_dataset(DatasetArn)
}
\arguments{
\item{DatasetArn}{[required] The Amazon Resource Name (ARN) of the dataset to delete.}
}
\description{
Deletes an Amazon Forecast dataset that was created using the
CreateDataset operation. You can only delete datasets that have a status
of \code{ACTIVE} or \code{CREATE_FAILED}. To get the status use the
DescribeDataset operation.
}
\details{
Forecast does not automatically update any dataset groups that contain
the deleted dataset. In order to update the dataset group, use the
operation, omitting the deleted dataset\'s ARN.
}
\section{Request syntax}{
\preformatted{svc$delete_dataset(
  DatasetArn = "string"
)
}
}

\keyword{internal}
