% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_dataset_group}
\alias{forecastservice_delete_dataset_group}
\title{Deletes a dataset group created using the CreateDatasetGroup operation}
\usage{
forecastservice_delete_dataset_group(DatasetGroupArn)
}
\arguments{
\item{DatasetGroupArn}{[required] The Amazon Resource Name (ARN) of the dataset group to delete.}
}
\description{
Deletes a dataset group created using the CreateDatasetGroup operation.
You can only delete dataset groups that have a status of \code{ACTIVE},
\code{CREATE_FAILED}, or \code{UPDATE_FAILED}. To get the status, use the
DescribeDatasetGroup operation.

This operation deletes only the dataset group, not the datasets in the
group.
}
\section{Request syntax}{
\preformatted{svc$delete_dataset_group(
  DatasetGroupArn = "string"
)
}
}

\keyword{internal}
