% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_forecast}
\alias{forecastservice_delete_forecast}
\title{Deletes a forecast created using the CreateForecast operation}
\usage{
forecastservice_delete_forecast(ForecastArn)
}
\arguments{
\item{ForecastArn}{[required] The Amazon Resource Name (ARN) of the forecast to delete.}
}
\description{
Deletes a forecast created using the CreateForecast operation. You can
delete only forecasts that have a status of \code{ACTIVE} or \code{CREATE_FAILED}.
To get the status, use the DescribeForecast operation.
}
\details{
You can\'t delete a forecast while it is being exported. After a
forecast is deleted, you can no longer query the forecast.
}
\section{Request syntax}{
\preformatted{svc$delete_forecast(
  ForecastArn = "string"
)
}
}

\keyword{internal}
