% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_dataset}
\alias{forecastservice_describe_dataset}
\title{Describes an Amazon Forecast dataset created using the CreateDataset
operation}
\usage{
forecastservice_describe_dataset(DatasetArn)
}
\arguments{
\item{DatasetArn}{[required] The Amazon Resource Name (ARN) of the dataset.}
}
\description{
Describes an Amazon Forecast dataset created using the CreateDataset
operation.

In addition to listing the parameters specified in the \code{CreateDataset}
request, this operation includes the following dataset properties:
\itemize{
\item \code{CreationTime}
\item \code{LastModificationTime}
\item \code{Status}
}
}
\section{Request syntax}{
\preformatted{svc$describe_dataset(
  DatasetArn = "string"
)
}
}

\keyword{internal}
