% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_dataset_group}
\alias{forecastservice_describe_dataset_group}
\title{Describes a dataset group created using the CreateDatasetGroup operation}
\usage{
forecastservice_describe_dataset_group(DatasetGroupArn)
}
\arguments{
\item{DatasetGroupArn}{[required] The Amazon Resource Name (ARN) of the dataset group.}
}
\description{
Describes a dataset group created using the
\code{\link[=forecastservice_create_dataset_group]{create_dataset_group}}
operation.

In addition to listing the parameters provided in the
\code{\link[=forecastservice_create_dataset_group]{create_dataset_group}} request,
this operation includes the following properties:
\itemize{
\item \code{DatasetArns} - The datasets belonging to the group.
\item \code{CreationTime}
\item \code{LastModificationTime}
\item \code{Status}
}
}
\section{Request syntax}{
\preformatted{svc$describe_dataset_group(
  DatasetGroupArn = "string"
)
}
}

\keyword{internal}
