% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_dataset_import_job}
\alias{forecastservice_describe_dataset_import_job}
\title{Describes a dataset import job created using the CreateDatasetImportJob
operation}
\usage{
forecastservice_describe_dataset_import_job(DatasetImportJobArn)
}
\arguments{
\item{DatasetImportJobArn}{[required] The Amazon Resource Name (ARN) of the dataset import job.}
}
\description{
Describes a dataset import job created using the
\code{\link[=forecastservice_create_dataset_import_job]{create_dataset_import_job}}
operation.

In addition to listing the parameters provided in the
\code{\link[=forecastservice_create_dataset_import_job]{create_dataset_import_job}}
request, this operation includes the following properties:
\itemize{
\item \code{CreationTime}
\item \code{LastModificationTime}
\item \code{DataSize}
\item \code{FieldStatistics}
\item \code{Status}
\item \code{Message} - If an error occurred, information about the error.
}
}
\section{Request syntax}{
\preformatted{svc$describe_dataset_import_job(
  DatasetImportJobArn = "string"
)
}
}

\keyword{internal}
