% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_forecast}
\alias{forecastservice_describe_forecast}
\title{Describes a forecast created using the CreateForecast operation}
\usage{
forecastservice_describe_forecast(ForecastArn)
}
\arguments{
\item{ForecastArn}{[required] The Amazon Resource Name (ARN) of the forecast.}
}
\description{
Describes a forecast created using the CreateForecast operation.

In addition to listing the properties provided in the \code{CreateForecast}
request, this operation lists the following properties:
\itemize{
\item \code{DatasetGroupArn} - The dataset group that provided the training
data.
\item \code{CreationTime}
\item \code{LastModificationTime}
\item \code{Status}
\item \code{Message} - If an error occurred, information about the error.
}
}
\section{Request syntax}{
\preformatted{svc$describe_forecast(
  ForecastArn = "string"
)
}
}

\keyword{internal}
