% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_forecast_export_job}
\alias{forecastservice_describe_forecast_export_job}
\title{Describes a forecast export job created using the
CreateForecastExportJob operation}
\usage{
forecastservice_describe_forecast_export_job(ForecastExportJobArn)
}
\arguments{
\item{ForecastExportJobArn}{[required] The Amazon Resource Name (ARN) of the forecast export job.}
}
\description{
Describes a forecast export job created using the
\code{\link[=forecastservice_create_forecast_export_job]{create_forecast_export_job}}
operation.

In addition to listing the properties provided by the user in the
\code{\link[=forecastservice_create_forecast_export_job]{create_forecast_export_job}}
request, this operation lists the following properties:
\itemize{
\item \code{CreationTime}
\item \code{LastModificationTime}
\item \code{Status}
\item \code{Message} - If an error occurred, information about the error.
}
}
\section{Request syntax}{
\preformatted{svc$describe_forecast_export_job(
  ForecastExportJobArn = "string"
)
}
}

\keyword{internal}
