% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_predictor}
\alias{forecastservice_describe_predictor}
\title{Describes a predictor created using the CreatePredictor operation}
\usage{
forecastservice_describe_predictor(PredictorArn)
}
\arguments{
\item{PredictorArn}{[required] The Amazon Resource Name (ARN) of the predictor that you want
information about.}
}
\description{
Describes a predictor created using the CreatePredictor operation.

In addition to listing the properties provided in the \code{CreatePredictor}
request, this operation lists the following properties:
\itemize{
\item \code{DatasetImportJobArns} - The dataset import jobs used to import
training data.
\item \code{AutoMLAlgorithmArns} - If AutoML is performed, the algorithms that
were evaluated.
\item \code{CreationTime}
\item \code{LastModificationTime}
\item \code{Status}
\item \code{Message} - If an error occurred, information about the error.
}
}
\section{Request syntax}{
\preformatted{svc$describe_predictor(
  PredictorArn = "string"
)
}
}

\keyword{internal}
