% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_predictor_backtest_export_job}
\alias{forecastservice_describe_predictor_backtest_export_job}
\title{Describes a predictor backtest export job created using the
CreatePredictorBacktestExportJob operation}
\usage{
forecastservice_describe_predictor_backtest_export_job(
  PredictorBacktestExportJobArn)
}
\arguments{
\item{PredictorBacktestExportJobArn}{[required] The Amazon Resource Name (ARN) of the predictor backtest export job.}
}
\description{
Describes a predictor backtest export job created using the
\code{\link[=forecastservice_create_predictor_backtest_export_job]{create_predictor_backtest_export_job}}
operation.

In addition to listing the properties provided by the user in the
\code{\link[=forecastservice_create_predictor_backtest_export_job]{create_predictor_backtest_export_job}}
request, this operation lists the following properties:
\itemize{
\item \code{CreationTime}
\item \code{LastModificationTime}
\item \code{Status}
\item \code{Message} (if an error occurred)
}
}
\section{Request syntax}{
\preformatted{svc$describe_predictor_backtest_export_job(
  PredictorBacktestExportJobArn = "string"
)
}
}

\keyword{internal}
