% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_get_accuracy_metrics}
\alias{forecastservice_get_accuracy_metrics}
\title{Provides metrics on the accuracy of the models that were trained by the
CreatePredictor operation}
\usage{
forecastservice_get_accuracy_metrics(PredictorArn)
}
\arguments{
\item{PredictorArn}{[required] The Amazon Resource Name (ARN) of the predictor to get metrics for.}
}
\description{
Provides metrics on the accuracy of the models that were trained by the
CreatePredictor operation. Use metrics to see how well the model
performed and to decide whether to use the predictor to generate a
forecast. For more information, see metrics.
}
\details{
This operation generates metrics for each backtest window that was
evaluated. The number of backtest windows (\code{NumberOfBacktestWindows}) is
specified using the EvaluationParameters object, which is optionally
included in the \code{CreatePredictor} request. If \code{NumberOfBacktestWindows}
isn\'t specified, the number defaults to one.

The parameters of the \code{filling} method determine which items contribute
to the metrics. If you want all items to contribute, specify \code{zero}. If
you want only those items that have complete data in the range being
evaluated to contribute, specify \code{nan}. For more information, see
FeaturizationMethod.

Before you can get accuracy metrics, the \code{Status} of the predictor must
be \code{ACTIVE}, signifying that training has completed. To get the status,
use the DescribePredictor operation.
}
\section{Request syntax}{
\preformatted{svc$get_accuracy_metrics(
  PredictorArn = "string"
)
}
}

\keyword{internal}
