% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_list_datasets}
\alias{forecastservice_list_datasets}
\title{Returns a list of datasets created using the CreateDataset operation}
\usage{
forecastservice_list_datasets(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{If the result of the previous request was truncated, the response
includes a \code{NextToken}. To retrieve the next set of results, use the
token in the next request. Tokens expire after 24 hours.}

\item{MaxResults}{The number of items to return in the response.}
}
\description{
Returns a list of datasets created using the
\code{\link[=forecastservice_create_dataset]{create_dataset}} operation. For each
dataset, a summary of its properties, including its Amazon Resource Name
(ARN), is returned. To retrieve the complete set of properties, use the
ARN with the \code{\link[=forecastservice_describe_dataset]{describe_dataset}}
operation.
}
\section{Request syntax}{
\preformatted{svc$list_datasets(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
