% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_list_predictor_backtest_export_jobs}
\alias{forecastservice_list_predictor_backtest_export_jobs}
\title{Returns a list of predictor backtest export jobs created using the
CreatePredictorBacktestExportJob operation}
\usage{
forecastservice_list_predictor_backtest_export_jobs(NextToken,
  MaxResults, Filters)
}
\arguments{
\item{NextToken}{If the result of the previous request was truncated, the response
includes a NextToken. To retrieve the next set of results, use the token
in the next request. Tokens expire after 24 hours.}

\item{MaxResults}{The number of items to return in the response.}

\item{Filters}{An array of filters. For each filter, provide a condition and a match
statement. The condition is either \code{IS} or \code{IS_NOT}, which specifies
whether to include or exclude the predictor backtest export jobs that
match the statement from the list. The match statement consists of a key
and a value.

\strong{Filter properties}
\itemize{
\item \code{Condition} - The condition to apply. Valid values are \code{IS} and
\code{IS_NOT}. To include the predictor backtest export jobs that match
the statement, specify \code{IS}. To exclude matching predictor backtest
export jobs, specify \code{IS_NOT}.
\item \code{Key} - The name of the parameter to filter on. Valid values are
\code{PredictorBacktestExportJobArn} and \code{Status}.
\item \code{Value} - The value to match.
}}
}
\description{
Returns a list of predictor backtest export jobs created using the
CreatePredictorBacktestExportJob operation. This operation returns a
summary for each backtest export job. You can filter the list using an
array of Filter objects.
}
\details{
To retrieve the complete set of properties for a particular backtest
export job, use the ARN with the DescribePredictorBacktestExportJob
operation.
}
\section{Request syntax}{
\preformatted{svc$list_predictor_backtest_export_jobs(
  NextToken = "string",
  MaxResults = 123,
  Filters = list(
    list(
      Key = "string",
      Value = "string",
      Condition = "IS"|"IS_NOT"
    )
  )
)
}
}

\keyword{internal}
