% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_list_predictors}
\alias{forecastservice_list_predictors}
\title{Returns a list of predictors created using the CreatePredictor operation}
\usage{
forecastservice_list_predictors(NextToken, MaxResults, Filters)
}
\arguments{
\item{NextToken}{If the result of the previous request was truncated, the response
includes a \code{NextToken}. To retrieve the next set of results, use the
token in the next request. Tokens expire after 24 hours.}

\item{MaxResults}{The number of items to return in the response.}

\item{Filters}{An array of filters. For each filter, you provide a condition and a
match statement. The condition is either \code{IS} or \code{IS_NOT}, which
specifies whether to include or exclude the predictors that match the
statement from the list, respectively. The match statement consists of a
key and a value.

\strong{Filter properties}
\itemize{
\item \code{Condition} - The condition to apply. Valid values are \code{IS} and
\code{IS_NOT}. To include the predictors that match the statement,
specify \code{IS}. To exclude matching predictors, specify \code{IS_NOT}.
\item \code{Key} - The name of the parameter to filter on. Valid values are
\code{DatasetGroupArn} and \code{Status}.
\item \code{Value} - The value to match.
}

For example, to list all predictors whose status is ACTIVE, you would
specify:

\verb{"Filters": \\\[ \\\{ "Condition": "IS", "Key": "Status", "Value": "ACTIVE" \\\} \\\]}}
}
\description{
Returns a list of predictors created using the
\code{\link[=forecastservice_create_predictor]{create_predictor}} operation. For
each predictor, this operation returns a summary of its properties,
including its Amazon Resource Name (ARN). You can retrieve the complete
set of properties by using the ARN with the
\code{\link[=forecastservice_describe_predictor]{describe_predictor}} operation.
You can filter the list using an array of Filter objects.
}
\section{Request syntax}{
\preformatted{svc$list_predictors(
  NextToken = "string",
  MaxResults = 123,
  Filters = list(
    list(
      Key = "string",
      Value = "string",
      Condition = "IS"|"IS_NOT"
    )
  )
)
}
}

\keyword{internal}
