% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_list_tags_for_resource}
\alias{forecastservice_list_tags_for_resource}
\title{Lists the tags for an Amazon Forecast resource}
\usage{
forecastservice_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource for which to
list the tags. Currently, the supported resources are Forecast dataset
groups, datasets, dataset import jobs, predictors, forecasts, and
forecast export jobs.}
}
\description{
Lists the tags for an Amazon Forecast resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
