% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_batch_create_variable}
\alias{frauddetector_batch_create_variable}
\title{Creates a batch of variables}
\usage{
frauddetector_batch_create_variable(variableEntries, tags)
}
\arguments{
\item{variableEntries}{[required] The list of variables for the batch create variable request.}

\item{tags}{A collection of key and value pairs.}
}
\description{
Creates a batch of variables.
}
\section{Request syntax}{
\preformatted{svc$batch_create_variable(
  variableEntries = list(
    list(
      name = "string",
      dataType = "string",
      dataSource = "string",
      defaultValue = "string",
      description = "string",
      variableType = "string"
    )
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
