% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_create_variable}
\alias{frauddetector_create_variable}
\title{Creates a variable}
\usage{
frauddetector_create_variable(name, dataType, dataSource, defaultValue,
  description, variableType)
}
\arguments{
\item{name}{[required] The name of the variable.}

\item{dataType}{[required] The data type.}

\item{dataSource}{[required] The source of the data.}

\item{defaultValue}{[required] The default value for the variable when no value is received.}

\item{description}{The description.}

\item{variableType}{The variable type.}
}
\description{
Creates a variable.
}
\section{Request syntax}{
\preformatted{svc$create_variable(
  name = "string",
  dataType = "STRING"|"INTEGER"|"FLOAT"|"BOOLEAN",
  dataSource = "EVENT"|"MODEL_SCORE"|"EXTERNAL_MODEL_SCORE",
  defaultValue = "string",
  description = "string",
  variableType = "string"
)
}
}

\keyword{internal}
