% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_detector_version}
\alias{frauddetector_delete_detector_version}
\title{Deletes the detector version}
\usage{
frauddetector_delete_detector_version(detectorId, detectorVersionId)
}
\arguments{
\item{detectorId}{[required] The ID of the parent detector for the detector version to delete.}

\item{detectorVersionId}{[required] The ID of the detector version to delete.}
}
\description{
Deletes the detector version.
}
\section{Request syntax}{
\preformatted{svc$delete_detector_version(
  detectorId = "string",
  detectorVersionId = "string"
)
}
}

\keyword{internal}
