% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_entity_type}
\alias{frauddetector_delete_entity_type}
\title{Deletes an entity type}
\usage{
frauddetector_delete_entity_type(name)
}
\arguments{
\item{name}{[required] The name of the entity type to delete.}
}
\description{
Deletes an entity type.

You cannot delete an entity type that is included in an event type.

When you delete an entity type, Amazon Fraud Detector permanently
deletes that entity type from the evaluation history, and the data is no
longer stored in Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$delete_entity_type(
  name = "string"
)
}
}

\keyword{internal}
