% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_event_type}
\alias{frauddetector_delete_event_type}
\title{Deletes an event type}
\usage{
frauddetector_delete_event_type(name)
}
\arguments{
\item{name}{[required] The name of the event type to delete.}
}
\description{
Deletes an event type.

You cannot delete an event type that is used in a detector or a model.

When you delete an entity type, Amazon Fraud Detector permanently
deletes that entity type from the evaluation history, and the data is no
longer stored in Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$delete_event_type(
  name = "string"
)
}
}

\keyword{internal}
