% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_external_model}
\alias{frauddetector_delete_external_model}
\title{Removes a SageMaker model from Amazon Fraud Detector}
\usage{
frauddetector_delete_external_model(modelEndpoint)
}
\arguments{
\item{modelEndpoint}{[required] The endpoint of the Amazon Sagemaker model to delete.}
}
\description{
Removes a SageMaker model from Amazon Fraud Detector.

You can remove an Amazon SageMaker model if it is not associated with a
detector version. Removing a SageMaker model disconnects it from Amazon
Fraud Detector, but the model remains available in SageMaker.
}
\section{Request syntax}{
\preformatted{svc$delete_external_model(
  modelEndpoint = "string"
)
}
}

\keyword{internal}
