% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_label}
\alias{frauddetector_delete_label}
\title{Deletes a label}
\usage{
frauddetector_delete_label(name)
}
\arguments{
\item{name}{[required] The name of the label to delete.}
}
\description{
Deletes a label.

You cannot delete labels that are included in an event type in Amazon
Fraud Detector.

You cannot delete a label assigned to an event ID. You must first delete
the relevant event ID.

When you delete a label, Amazon Fraud Detector permanently deletes that
label from the evaluation history, and the data is no longer stored in
Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$delete_label(
  name = "string"
)
}
}

\keyword{internal}
