% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_model}
\alias{frauddetector_delete_model}
\title{Deletes a model}
\usage{
frauddetector_delete_model(modelId, modelType)
}
\arguments{
\item{modelId}{[required] The model ID of the model to delete.}

\item{modelType}{[required] The model type of the model to delete.}
}
\description{
Deletes a model.
}
\details{
You can delete models and model versions in Amazon Fraud Detector,
provided that they are not associated with a detector version.

When you delete a model, Amazon Fraud Detector permanently deletes that
model from the evaluation history, and the data is no longer stored in
Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$delete_model(
  modelId = "string",
  modelType = "ONLINE_FRAUD_INSIGHTS"
)
}
}

\keyword{internal}
