% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_model_version}
\alias{frauddetector_delete_model_version}
\title{Deletes a model version}
\usage{
frauddetector_delete_model_version(modelId, modelType,
  modelVersionNumber)
}
\arguments{
\item{modelId}{[required] The model ID of the model version to delete.}

\item{modelType}{[required] The model type of the model version to delete.}

\item{modelVersionNumber}{[required] The model version number of the model version to delete.}
}
\description{
Deletes a model version.

You can delete models and model versions in Amazon Fraud Detector,
provided that they are not associated with a detector version.

When you delete a model version, Amazon Fraud Detector permanently
deletes that model version from the evaluation history, and the data is
no longer stored in Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$delete_model_version(
  modelId = "string",
  modelType = "ONLINE_FRAUD_INSIGHTS",
  modelVersionNumber = "string"
)
}
}

\keyword{internal}
