% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_outcome}
\alias{frauddetector_delete_outcome}
\title{Deletes an outcome}
\usage{
frauddetector_delete_outcome(name)
}
\arguments{
\item{name}{[required] The name of the outcome to delete.}
}
\description{
Deletes an outcome.
}
\details{
You cannot delete an outcome that is used in a rule version.

When you delete an outcome, Amazon Fraud Detector permanently deletes
that outcome from the evaluation history, and the data is no longer
stored in Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$delete_outcome(
  name = "string"
)
}
}

\keyword{internal}
