% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_rule}
\alias{frauddetector_delete_rule}
\title{Deletes the rule}
\usage{
frauddetector_delete_rule(rule)
}
\arguments{
\item{rule}{[required]}
}
\description{
Deletes the rule. You cannot delete a rule if it is used by an \code{ACTIVE}
or \code{INACTIVE} detector version.
}
\details{
When you delete a rule, Amazon Fraud Detector permanently deletes that
rule from the evaluation history, and the data is no longer stored in
Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$delete_rule(
  rule = list(
    detectorId = "string",
    ruleId = "string",
    ruleVersion = "string"
  )
)
}
}

\keyword{internal}
