% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_rule_version}
\alias{frauddetector_delete_rule_version}
\title{Deletes the rule version}
\usage{
frauddetector_delete_rule_version(detectorId, ruleId, ruleVersion)
}
\arguments{
\item{detectorId}{[required] The ID of the detector that includes the rule version to delete.}

\item{ruleId}{[required] The rule ID of the rule version to delete.}

\item{ruleVersion}{[required] The rule version to delete.}
}
\description{
Deletes the rule version. You cannot delete a rule version if it is used
by an \code{ACTIVE} or \code{INACTIVE} detector version.
}
\section{Request syntax}{
\preformatted{svc$delete_rule_version(
  detectorId = "string",
  ruleId = "string",
  ruleVersion = "string"
)
}
}

\keyword{internal}
