% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_variable}
\alias{frauddetector_delete_variable}
\title{Deletes a variable}
\usage{
frauddetector_delete_variable(name)
}
\arguments{
\item{name}{[required] The name of the variable to delete.}
}
\description{
Deletes a variable.
}
\details{
You can't delete variables that are included in an event type in Amazon
Fraud Detector.

Amazon Fraud Detector automatically deletes model output variables and
SageMaker model output variables when you delete the model. You can't
delete these variables manually.

When you delete a variable, Amazon Fraud Detector permanently deletes
that variable from the evaluation history, and the data is no longer
stored in Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$delete_variable(
  name = "string"
)
}
}

\keyword{internal}
