% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_describe_model_versions}
\alias{frauddetector_describe_model_versions}
\title{Gets all of the model versions for the specified model type or for the
specified model type and model ID}
\usage{
frauddetector_describe_model_versions(modelId, modelVersionNumber,
  modelType, nextToken, maxResults)
}
\arguments{
\item{modelId}{The model ID.}

\item{modelVersionNumber}{The model version number.}

\item{modelType}{The model type.}

\item{nextToken}{The next token from the previous results.}

\item{maxResults}{The maximum number of results to return.}
}
\description{
Gets all of the model versions for the specified model type or for the
specified model type and model ID. You can also get details for a
single, specified model version.
}
\section{Request syntax}{
\preformatted{svc$describe_model_versions(
  modelId = "string",
  modelVersionNumber = "string",
  modelType = "ONLINE_FRAUD_INSIGHTS",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
