% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_detectors}
\alias{frauddetector_get_detectors}
\title{Gets all of detectors}
\usage{
frauddetector_get_detectors(detectorId, nextToken, maxResults)
}
\arguments{
\item{detectorId}{The detector ID.}

\item{nextToken}{The next token for the subsequent request.}

\item{maxResults}{The maximum number of objects to return for the request.}
}
\description{
Gets all of detectors. This is a paginated API. If you provide a null
\code{maxSizePerPage}, this actions retrieves a maximum of 10 records per
page. If you provide a \code{maxSizePerPage}, the value must be between 5 and
10. To get the next page results, provide the pagination token from the
\code{GetEventTypesResponse} as part of your request. A null pagination token
fetches the records from the beginning.
}
\section{Request syntax}{
\preformatted{svc$get_detectors(
  detectorId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
