% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_entity_types}
\alias{frauddetector_get_entity_types}
\title{Gets all entity types or a specific entity type if a name is specified}
\usage{
frauddetector_get_entity_types(name, nextToken, maxResults)
}
\arguments{
\item{name}{The name.}

\item{nextToken}{The next token for the subsequent request.}

\item{maxResults}{The maximum number of objects to return for the request.}
}
\description{
Gets all entity types or a specific entity type if a name is specified.
This is a paginated API. If you provide a null \code{maxResults}, this action
retrieves a maximum of 10 records per page. If you provide a
\code{maxResults}, the value must be between 5 and 10. To get the next page
results, provide the pagination token from the \code{GetEntityTypesResponse}
as part of your request. A null pagination token fetches the records
from the beginning.
}
\section{Request syntax}{
\preformatted{svc$get_entity_types(
  name = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
