% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_event_prediction}
\alias{frauddetector_get_event_prediction}
\title{Evaluates an event against a detector version}
\usage{
frauddetector_get_event_prediction(detectorId, detectorVersionId,
  eventId, eventTypeName, entities, eventTimestamp, eventVariables,
  externalModelEndpointDataBlobs)
}
\arguments{
\item{detectorId}{[required] The detector ID.}

\item{detectorVersionId}{The detector version ID.}

\item{eventId}{[required] The unique ID used to identify the event.}

\item{eventTypeName}{[required] The event type associated with the detector specified for the
prediction.}

\item{entities}{[required] The entity type (associated with the detector's event type) and specific
entity ID representing who performed the event. If an entity id is not
available, use "UNKNOWN."}

\item{eventTimestamp}{[required] Timestamp that defines when the event under evaluation occurred.}

\item{eventVariables}{[required] Names of the event type's variables you defined in Amazon Fraud Detector
to represent data elements and their corresponding values for the event
you are sending for evaluation.}

\item{externalModelEndpointDataBlobs}{The Amazon SageMaker model endpoint input data blobs.}
}
\description{
Evaluates an event against a detector version. If a version ID is not
provided, the detector’s (\code{ACTIVE}) version is used.
}
\section{Request syntax}{
\preformatted{svc$get_event_prediction(
  detectorId = "string",
  detectorVersionId = "string",
  eventId = "string",
  eventTypeName = "string",
  entities = list(
    list(
      entityType = "string",
      entityId = "string"
    )
  ),
  eventTimestamp = "string",
  eventVariables = list(
    "string"
  ),
  externalModelEndpointDataBlobs = list(
    list(
      byteBuffer = raw,
      contentType = "string"
    )
  )
)
}
}

\keyword{internal}
