% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_kms_encryption_key}
\alias{frauddetector_get_kms_encryption_key}
\title{Gets the encryption key if a Key Management Service (KMS) customer
master key (CMK) has been specified to be used to encrypt content in
Amazon Fraud Detector}
\usage{
frauddetector_get_kms_encryption_key()
}
\description{
Gets the encryption key if a Key Management Service (KMS) customer
master key (CMK) has been specified to be used to encrypt content in
Amazon Fraud Detector.
}
\keyword{internal}
