% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_models}
\alias{frauddetector_get_models}
\title{Gets one or more models}
\usage{
frauddetector_get_models(modelId, modelType, nextToken, maxResults)
}
\arguments{
\item{modelId}{The model ID.}

\item{modelType}{The model type.}

\item{nextToken}{The next token for the subsequent request.}

\item{maxResults}{The maximum number of objects to return for the request.}
}
\description{
Gets one or more models. Gets all models for the AWS account if no model
type and no model id provided. Gets all models for the AWS account and
model type, if the model type is specified but model id is not provided.
Gets a specific model if (model type, model id) tuple is specified.

This is a paginated API. If you provide a null \code{maxResults}, this action
retrieves a maximum of 10 records per page. If you provide a
\code{maxResults}, the value must be between 1 and 10. To get the next page
results, provide the pagination token from the response as part of your
request. A null pagination token fetches the records from the beginning.
}
\section{Request syntax}{
\preformatted{svc$get_models(
  modelId = "string",
  modelType = "ONLINE_FRAUD_INSIGHTS",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
