% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_prediction}
\alias{frauddetector_get_prediction}
\title{Evaluates an event against a detector version}
\usage{
frauddetector_get_prediction(detectorId, detectorVersionId, eventId,
  eventAttributes, externalModelEndpointDataBlobs)
}
\arguments{
\item{detectorId}{[required] The detector ID.}

\item{detectorVersionId}{The detector version ID.}

\item{eventId}{[required] The unique ID used to identify the event.}

\item{eventAttributes}{Names of variables you defined in Amazon Fraud Detector to represent
event data elements and their corresponding values for the event you are
sending for evaluation.}

\item{externalModelEndpointDataBlobs}{The Amazon SageMaker model endpoint input data blobs.}
}
\description{
Evaluates an event against a detector version. If a version ID is not
provided, the detector’s (\code{ACTIVE}) version is used.
}
\section{Request syntax}{
\preformatted{svc$get_prediction(
  detectorId = "string",
  detectorVersionId = "string",
  eventId = "string",
  eventAttributes = list(
    "string"
  ),
  externalModelEndpointDataBlobs = list(
    list(
      byteBuffer = raw,
      contentType = "string"
    )
  )
)
}
}

\keyword{internal}
