% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_rules}
\alias{frauddetector_get_rules}
\title{Gets all rules available for the specified detector}
\usage{
frauddetector_get_rules(ruleId, detectorId, ruleVersion, nextToken,
  maxResults)
}
\arguments{
\item{ruleId}{The rule ID.}

\item{detectorId}{[required] The detector ID.}

\item{ruleVersion}{The rule version.}

\item{nextToken}{The next page token.}

\item{maxResults}{The maximum number of rules to return for the request.}
}
\description{
Gets all rules available for the specified detector.
}
\section{Request syntax}{
\preformatted{svc$get_rules(
  ruleId = "string",
  detectorId = "string",
  ruleVersion = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
