% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_put_detector}
\alias{frauddetector_put_detector}
\title{Creates or updates a detector}
\usage{
frauddetector_put_detector(detectorId, description)
}
\arguments{
\item{detectorId}{[required] The detector ID.}

\item{description}{The description of the detector.}
}
\description{
Creates or updates a detector.
}
\section{Request syntax}{
\preformatted{svc$put_detector(
  detectorId = "string",
  description = "string"
)
}
}

\keyword{internal}
