% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_put_external_model}
\alias{frauddetector_put_external_model}
\title{Creates or updates an Amazon SageMaker model endpoint}
\usage{
frauddetector_put_external_model(modelEndpoint, modelSource, role,
  inputConfiguration, outputConfiguration, modelEndpointStatus)
}
\arguments{
\item{modelEndpoint}{[required] The model endpoints name.}

\item{modelSource}{[required] The source of the model.}

\item{role}{[required] The IAM role used to invoke the model endpoint.}

\item{inputConfiguration}{[required] The model endpoint input configuration.}

\item{outputConfiguration}{[required] The model endpoint output configuration.}

\item{modelEndpointStatus}{[required] The model endpoint's status in Amazon Fraud Detector.}
}
\description{
Creates or updates an Amazon SageMaker model endpoint. You can also use
this action to update the configuration of the model endpoint, including
the IAM role and/or the mapped variables.
}
\section{Request syntax}{
\preformatted{svc$put_external_model(
  modelEndpoint = "string",
  modelSource = "SAGEMAKER",
  role = list(
    arn = "string",
    name = "string"
  ),
  inputConfiguration = list(
    format = "TEXT_CSV"|"APPLICATION_JSON",
    isOpaque = TRUE|FALSE,
    jsonInputTemplate = "string",
    csvInputTemplate = "string"
  ),
  outputConfiguration = list(
    format = "TEXT_CSV"|"APPLICATION_JSONLINES",
    jsonKeyToVariableMap = list(
      "string"
    ),
    csvIndexToVariableMap = list(
      "string"
    )
  ),
  modelEndpointStatus = "ASSOCIATED"|"DISSOCIATED"
)
}
}

\keyword{internal}
