% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_put_label}
\alias{frauddetector_put_label}
\title{Creates or updates label}
\usage{
frauddetector_put_label(name, description, tags)
}
\arguments{
\item{name}{[required] The label name.}

\item{description}{The label description.}

\item{tags}{}
}
\description{
Creates or updates label. A label classifies an event as fraudulent or
legitimate. Labels are associated with event types and used to train
supervised machine learning models in Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$put_label(
  name = "string",
  description = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
