% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_put_model}
\alias{frauddetector_put_model}
\title{Creates or updates a model}
\usage{
frauddetector_put_model(modelId, modelType, description,
  trainingDataSource, modelVariables, labelSchema)
}
\arguments{
\item{modelId}{[required] The model ID.}

\item{modelType}{[required] The model type.}

\item{description}{The model description.}

\item{trainingDataSource}{[required] The training data source location in Amazon S3.}

\item{modelVariables}{[required] The model input variables.}

\item{labelSchema}{[required] The label schema.}
}
\description{
Creates or updates a model.
}
\section{Request syntax}{
\preformatted{svc$put_model(
  modelId = "string",
  modelType = "ONLINE_FRAUD_INSIGHTS",
  description = "string",
  trainingDataSource = list(
    dataLocation = "string",
    dataAccessRoleArn = "string"
  ),
  modelVariables = list(
    list(
      name = "string",
      index = 123
    )
  ),
  labelSchema = list(
    labelKey = "string",
    labelMapper = list(
      list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
