% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_tag_resource}
\alias{frauddetector_tag_resource}
\title{Assigns tags to a resource}
\usage{
frauddetector_tag_resource(resourceARN, tags)
}
\arguments{
\item{resourceARN}{[required] The resource ARN.}

\item{tags}{[required] The tags to assign to the resource.}
}
\description{
Assigns tags to a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceARN = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
