% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_model}
\alias{frauddetector_update_model}
\title{Updates a model}
\usage{
frauddetector_update_model(modelId, modelType, description)
}
\arguments{
\item{modelId}{[required] The model ID.}

\item{modelType}{[required] The model type.}

\item{description}{The new model description.}
}
\description{
Updates a model. You can update the description attribute using this
action.
}
\section{Request syntax}{
\preformatted{svc$update_model(
  modelId = "string",
  modelType = "ONLINE_FRAUD_INSIGHTS",
  description = "string"
)
}
}

\keyword{internal}
