% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_model_version}
\alias{frauddetector_update_model_version}
\title{Updates a model version}
\usage{
frauddetector_update_model_version(modelId, modelType,
  modelVersionNumber, description, status)
}
\arguments{
\item{modelId}{[required] The model ID.}

\item{modelType}{[required] The model type.}

\item{modelVersionNumber}{[required] The model version.}

\item{description}{[required] The model description.}

\item{status}{[required] The new model status.}
}
\description{
Updates a model version. You can update the description and status
attributes using this action. You can perform the following status
updates:
}
\details{
\enumerate{
\item Change the \code{TRAINING_COMPLETE} status to \code{ACTIVE}
\item Change \code{ACTIVE} back to \code{TRAINING_COMPLETE}
}
}
\section{Request syntax}{
\preformatted{svc$update_model_version(
  modelId = "string",
  modelType = "ONLINE_FRAUD_INSIGHTS",
  modelVersionNumber = "string",
  description = "string",
  status = "TRAINING_IN_PROGRESS"|"TRAINING_COMPLETE"|"ACTIVATE_REQUESTED"|"ACTIVATE_IN_PROGRESS"|"ACTIVE"|"INACTIVATE_IN_PROGRESS"|"INACTIVE"|"ERROR"
)
}
}

\keyword{internal}
