% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_rule_metadata}
\alias{frauddetector_update_rule_metadata}
\title{Updates a rule's metadata}
\usage{
frauddetector_update_rule_metadata(rule, description)
}
\arguments{
\item{rule}{[required] The rule to update.}

\item{description}{[required] The rule description.}
}
\description{
Updates a rule's metadata.
}
\section{Request syntax}{
\preformatted{svc$update_rule_metadata(
  rule = list(
    detectorId = "string",
    ruleId = "string",
    ruleVersion = "string"
  ),
  description = "string"
)
}
}

\keyword{internal}
