% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_variable}
\alias{frauddetector_update_variable}
\title{Updates a variable}
\usage{
frauddetector_update_variable(name, defaultValue, description,
  variableType)
}
\arguments{
\item{name}{[required] The name of the variable.}

\item{defaultValue}{The new default value of the variable.}

\item{description}{The new description.}

\item{variableType}{The variable type. For more information see \href{https://docs.aws.amazon.com/frauddetector/latest/ug/create-a-variable.html#variable-types}{Variable types}.}
}
\description{
Updates a variable.
}
\section{Request syntax}{
\preformatted{svc$update_variable(
  name = "string",
  defaultValue = "string",
  description = "string",
  variableType = "string"
)
}
}

\keyword{internal}
