% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_service.R
\name{gamelift}
\alias{gamelift}
\title{Amazon GameLift}
\usage{
gamelift()
}
\description{
Amazon GameLift Service

Amazon GameLift is a managed service for developers who need a scalable,
dedicated server solution for their multiplayer games. Use Amazon
GameLift for these tasks: (1) set up computing resources and deploy your
game servers, (2) run game sessions and get players into games, (3)
automatically scale your resources to meet player demand and manage
costs, and (4) track in-depth metrics on game server performance and
player usage.

The Amazon GameLift service API includes two important function sets:
\itemize{
\item \strong{Manage game sessions and player access} \-- Retrieve information
on available game sessions; create new game sessions; send player
requests to join a game session.
\item \strong{Configure and manage game server resources} \-- Manage builds,
fleets, queues, and aliases; set auto-scaling policies; retrieve
logs and metrics.
}

This reference guide describes the low-level service API for Amazon
GameLift. You can use the API functionality with these tools:
\itemize{
\item The Amazon Web Services software development kit (\href{http://aws.amazon.com/tools/#sdk}{AWS SDK}) is available in \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-supported.html#gamelift-supported-clients}{multiple languages}
including C++ and C\#. Use the SDK to access the API
programmatically from an application, such as a game client.
\item The \href{https://docs.aws.amazon.com/cli/latest/userguide/}{AWS command-line interface} (CLI)
tool is primarily useful for handling administrative actions, such
as setting up and managing Amazon GameLift settings and resources.
You can use the AWS CLI to manage all of your AWS services.
\item The \href{https://console.aws.amazon.com/gamelift/home}{AWS Management Console} for Amazon
GameLift provides a web interface to manage your Amazon GameLift
settings and resources. The console includes a dashboard for
tracking key resources, including builds and fleets, and displays
usage and performance metrics for your games as customizable graphs.
\item Amazon GameLift Local is a tool for testing your game's integration
with Amazon GameLift before deploying it on the service. This tools
supports a subset of key API actions, which can be called from
either the AWS CLI or programmatically. See \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/integration-testing-local.html}{Testing an Integration}.
}

\strong{Learn more}
\itemize{
\item \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/}{Developer Guide}
\-- Read about Amazon GameLift features and how to use them.
\item \href{https://gamedev.amazon.com/forums/tutorials}{Tutorials} \-- Get
started fast with walkthroughs and sample projects.
\item \href{https://gamedev.amazon.com/blogs/gamedev/}{GameDev Blog} \-- Stay
up to date with new features and techniques.
\item GameDev Forums \-- Connect with the GameDev community.
\item \href{https://aws.amazon.com/releasenotes/Amazon-GameLift/}{Release notes} and
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/doc-history.html}{document history}
\-- Stay current with updates to the Amazon GameLift service, SDKs,
and documentation.
}

\strong{API SUMMARY}

This list offers a functional overview of the Amazon GameLift service
API.

\strong{Managing Games and Players}

Use these actions to start new game sessions, find existing game
sessions, track game session status and other information, and enable
player access to game sessions.
\itemize{
\item \strong{Discover existing game sessions}
\itemize{
\item SearchGameSessions \-- Retrieve all available game sessions or
search for game sessions that match a set of criteria.
}
\item \strong{Start new game sessions}
\itemize{
\item Start new games with Queues to find the best available hosting
resources across multiple regions, minimize player latency, and
balance game session activity for efficiency and cost
effectiveness.
\itemize{
\item StartGameSessionPlacement \-- Request a new game session
placement and add one or more players to it.
\item DescribeGameSessionPlacement \-- Get details on a placement
request, including status.
\item StopGameSessionPlacement \-- Cancel a placement request.
}
\item CreateGameSession \-- Start a new game session on a specific
fleet. \emph{Available in Amazon GameLift Local.}
}
\item \strong{Match players to game sessions with FlexMatch matchmaking}
\itemize{
\item StartMatchmaking \-- Request matchmaking for one players or a
group who want to play together.
\item StartMatchBackfill - Request additional player matches to fill
empty slots in an existing game session.
\item DescribeMatchmaking \-- Get details on a matchmaking request,
including status.
\item AcceptMatch \-- Register that a player accepts a proposed match,
for matches that require player acceptance.
\item StopMatchmaking \-- Cancel a matchmaking request.
}
\item \strong{Manage game session data}
\itemize{
\item DescribeGameSessions \-- Retrieve metadata for one or more game
sessions, including length of time active and current player
count. \emph{Available in Amazon GameLift Local.}
\item DescribeGameSessionDetails \-- Retrieve metadata and the game
session protection setting for one or more game sessions.
\item UpdateGameSession \-- Change game session settings, such as
maximum player count and join policy.
\item GetGameSessionLogUrl \-- Get the location of saved logs for a
game session.
}
\item \strong{Manage player sessions}
\itemize{
\item CreatePlayerSession \-- Send a request for a player to join a
game session. \emph{Available in Amazon GameLift Local.}
\item CreatePlayerSessions \-- Send a request for multiple players to
join a game session. \emph{Available in Amazon GameLift Local.}
\item DescribePlayerSessions \-- Get details on player activity,
including status, playing time, and player data. \emph{Available in
Amazon GameLift Local.}
}
}

\strong{Setting Up and Managing Game Servers}

When setting up Amazon GameLift resources for your game, you first
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-build-intro.html}{create a game build}
and upload it to Amazon GameLift. You can then use these actions to
configure and manage a fleet of resources to run your game servers,
scale capacity to meet player demand, access performance and utilization
metrics, and more.
\itemize{
\item \strong{Manage game builds}
\itemize{
\item CreateBuild \-- Create a new build using files stored in an
Amazon S3 bucket. To create a build and upload files from a
local path, use the AWS CLI command \code{upload-build}.
\item ListBuilds \-- Get a list of all builds uploaded to a Amazon
GameLift region.
\item DescribeBuild \-- Retrieve information associated with a build.
\item UpdateBuild \-- Change build metadata, including build name and
version.
\item DeleteBuild \-- Remove a build from Amazon GameLift.
}
\item \strong{Manage fleets}
\itemize{
\item CreateFleet \-- Configure and activate a new fleet to run a
build's game servers.
\item ListFleets \-- Get a list of all fleet IDs in a Amazon GameLift
region (all statuses).
\item DeleteFleet \-- Terminate a fleet that is no longer running game
servers or hosting players.
\item View / update fleet configurations.
\itemize{
\item DescribeFleetAttributes / UpdateFleetAttributes \-- View or
change a fleet's metadata and settings for game session
protection and resource creation limits.
\item DescribeFleetPortSettings / UpdateFleetPortSettings \-- View
or change the inbound permissions (IP address and port
setting ranges) allowed for a fleet.
\item DescribeRuntimeConfiguration / UpdateRuntimeConfiguration
\-- View or change what server processes (and how many) to
run on each instance in a fleet.
}
}
\item \strong{Control fleet capacity}
\itemize{
\item DescribeEC2InstanceLimits \-- Retrieve maximum number of
instances allowed for the current AWS account and the current
usage level.
\item DescribeFleetCapacity / UpdateFleetCapacity \-- Retrieve the
capacity settings and the current number of instances in a
fleet; adjust fleet capacity settings to scale up or down.
\item Autoscale \-- Manage auto-scaling rules and apply them to a
fleet.
\itemize{
\item PutScalingPolicy \-- Create a new auto-scaling policy, or
update an existing one.
\item DescribeScalingPolicies \-- Retrieve an existing
auto-scaling policy.
\item DeleteScalingPolicy \-- Delete an auto-scaling policy and
stop it from affecting a fleet's capacity.
\item StartFleetActions \-- Restart a fleet's auto-scaling
policies.
\item StopFleetActions \-- Suspend a fleet's auto-scaling
policies.
}
}
\item \strong{Manage game session queues}
\itemize{
\item CreateGameSessionQueue \-- Create a queue for processing
requests for new game sessions.
\item DescribeGameSessionQueues \-- Retrieve game session queues
defined in a Amazon GameLift region.
\item UpdateGameSessionQueue \-- Change the configuration of a game
session queue.
\item DeleteGameSessionQueue \-- Remove a game session queue from the
region.
}
\item \strong{Manage FlexMatch resources}
\itemize{
\item CreateMatchmakingConfiguration \-- Create a matchmaking
configuration with instructions for building a player group and
placing in a new game session.
\item DescribeMatchmakingConfigurations \-- Retrieve matchmaking
configurations defined a Amazon GameLift region.
\item UpdateMatchmakingConfiguration \-- Change settings for
matchmaking configuration. queue.
\item DeleteMatchmakingConfiguration \-- Remove a matchmaking
configuration from the region.
\item CreateMatchmakingRuleSet \-- Create a set of rules to use when
searching for player matches.
\item DescribeMatchmakingRuleSets \-- Retrieve matchmaking rule sets
defined in a Amazon GameLift region.
\item ValidateMatchmakingRuleSet \-- Verify syntax for a set of
matchmaking rules.
\item DeleteMatchmakingRuleSet \-- Remove a matchmaking rule set from
the region.
}
\item \strong{Access fleet activity statistics}
\itemize{
\item DescribeFleetUtilization \-- Get current data on the number of
server processes, game sessions, and players currently active on
a fleet.
\item DescribeFleetEvents \-- Get a fleet's logged events for a
specified time span.
\item DescribeGameSessions \-- Retrieve metadata associated with one
or more game sessions, including length of time active and
current player count.
}
\item \strong{Remotely access an instance}
\itemize{
\item DescribeInstances \-- Get information on each instance in a
fleet, including instance ID, IP address, and status.
\item GetInstanceAccess \-- Request access credentials needed to
remotely connect to a specified instance in a fleet.
}
\item \strong{Manage fleet aliases}
\itemize{
\item CreateAlias \-- Define a new alias and optionally assign it to a
fleet.
\item ListAliases \-- Get all fleet aliases defined in a Amazon
GameLift region.
\item DescribeAlias \-- Retrieve information on an existing alias.
\item UpdateAlias \-- Change settings for a alias, such as redirecting
it from one fleet to another.
\item DeleteAlias \-- Remove an alias from the region.
\item ResolveAlias \-- Get the fleet ID that a specified alias points
to.
}
\item \strong{Manage VPC peering connections for fleets}
\itemize{
\item CreateVpcPeeringAuthorization \-- Authorize a peering connection
to one of your VPCs.
\item DescribeVpcPeeringAuthorizations \-- Retrieve valid peering
connection authorizations.
\item DeleteVpcPeeringAuthorization \-- Delete a peering connection
authorization.
\item CreateVpcPeeringConnection \-- Establish a peering connection
between the VPC for a Amazon GameLift fleet and one of your
VPCs.
\item DescribeVpcPeeringConnections \-- Retrieve information on active
or pending VPC peering connections with a Amazon GameLift fleet.
\item DeleteVpcPeeringConnection \-- Delete a VPC peering connection
with a Amazon GameLift fleet.
}
}
}
\section{Example}{
\preformatted{gamelift <- paws::gamelift()
gamelift$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=gamelift_accept_match]{accept_match} \tab Registers a player's acceptance or rejection of a proposed FlexMatch match \cr
\link[=gamelift_create_alias]{create_alias} \tab Creates an alias for a fleet \cr
\link[=gamelift_create_build]{create_build} \tab Creates a new Amazon GameLift build record for your game server binary files and points to the location of your game server build files in an Amazon Simple Storage Service (Amazon S3) location \cr
\link[=gamelift_create_fleet]{create_fleet} \tab Creates a new fleet to run your game servers \cr
\link[=gamelift_create_game_session]{create_game_session} \tab Creates a multiplayer game session for players \cr
\link[=gamelift_create_game_session_queue]{create_game_session_queue} \tab Establishes a new queue for processing requests to place new game sessions \cr
\link[=gamelift_create_matchmaking_configuration]{create_matchmaking_configuration} \tab Defines a new matchmaking configuration for use with FlexMatch \cr
\link[=gamelift_create_matchmaking_rule_set]{create_matchmaking_rule_set} \tab Creates a new rule set for FlexMatch matchmaking \cr
\link[=gamelift_create_player_session]{create_player_session} \tab Adds a player to a game session and creates a player session record \cr
\link[=gamelift_create_player_sessions]{create_player_sessions} \tab Adds a group of players to a game session \cr
\link[=gamelift_create_vpc_peering_authorization]{create_vpc_peering_authorization} \tab Requests authorization to create or delete a peer connection between the VPC for your Amazon GameLift fleet and a virtual private cloud (VPC) in your AWS account \cr
\link[=gamelift_create_vpc_peering_connection]{create_vpc_peering_connection} \tab Establishes a VPC peering connection between a virtual private cloud (VPC) in an AWS account with the VPC for your Amazon GameLift fleet \cr
\link[=gamelift_delete_alias]{delete_alias} \tab Deletes an alias \cr
\link[=gamelift_delete_build]{delete_build} \tab Deletes a build \cr
\link[=gamelift_delete_fleet]{delete_fleet} \tab Deletes everything related to a fleet \cr
\link[=gamelift_delete_game_session_queue]{delete_game_session_queue} \tab Deletes a game session queue \cr
\link[=gamelift_delete_matchmaking_configuration]{delete_matchmaking_configuration} \tab Permanently removes a FlexMatch matchmaking configuration \cr
\link[=gamelift_delete_matchmaking_rule_set]{delete_matchmaking_rule_set} \tab Deletes an existing matchmaking rule set \cr
\link[=gamelift_delete_scaling_policy]{delete_scaling_policy} \tab Deletes a fleet scaling policy \cr
\link[=gamelift_delete_vpc_peering_authorization]{delete_vpc_peering_authorization} \tab Cancels a pending VPC peering authorization for the specified VPC \cr
\link[=gamelift_delete_vpc_peering_connection]{delete_vpc_peering_connection} \tab Removes a VPC peering connection \cr
\link[=gamelift_describe_alias]{describe_alias} \tab Retrieves properties for an alias \cr
\link[=gamelift_describe_build]{describe_build} \tab Retrieves properties for a build \cr
\link[=gamelift_describe_ec2_instance_limits]{describe_ec2_instance_limits} \tab Retrieves the following information for the specified EC2 instance type: - maximum number of instances allowed per AWS account (service limit) - current usage level for the AWS account Service limits vary depending on region\cr
\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes} \tab Retrieves fleet properties, including metadata, status, and configuration, for one or more fleets \cr
\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity} \tab Retrieves the current status of fleet capacity for one or more fleets \cr
\link[=gamelift_describe_fleet_events]{describe_fleet_events} \tab Retrieves entries from the specified fleet's event log \cr
\link[=gamelift_describe_fleet_port_settings]{describe_fleet_port_settings} \tab Retrieves the inbound connection permissions for a fleet \cr
\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization} \tab Retrieves utilization statistics for one or more fleets \cr
\link[=gamelift_describe_game_session_details]{describe_game_session_details} \tab Retrieves properties, including the protection policy in force, for one or more game sessions \cr
\link[=gamelift_describe_game_session_placement]{describe_game_session_placement} \tab Retrieves properties and current status of a game session placement request \cr
\link[=gamelift_describe_game_session_queues]{describe_game_session_queues} \tab Retrieves the properties for one or more game session queues \cr
\link[=gamelift_describe_game_sessions]{describe_game_sessions} \tab Retrieves a set of one or more game sessions \cr
\link[=gamelift_describe_instances]{describe_instances} \tab Retrieves information about a fleet's instances, including instance IDs \cr
\link[=gamelift_describe_matchmaking]{describe_matchmaking} \tab Retrieves one or more matchmaking tickets \cr
\link[=gamelift_describe_matchmaking_configurations]{describe_matchmaking_configurations} \tab Retrieves the details of FlexMatch matchmaking configurations \cr
\link[=gamelift_describe_matchmaking_rule_sets]{describe_matchmaking_rule_sets} \tab Retrieves the details for FlexMatch matchmaking rule sets \cr
\link[=gamelift_describe_player_sessions]{describe_player_sessions} \tab Retrieves properties for one or more player sessions \cr
\link[=gamelift_describe_runtime_configuration]{describe_runtime_configuration} \tab Retrieves the current run-time configuration for the specified fleet \cr
\link[=gamelift_describe_scaling_policies]{describe_scaling_policies} \tab Retrieves all scaling policies applied to a fleet \cr
\link[=gamelift_describe_vpc_peering_authorizations]{describe_vpc_peering_authorizations} \tab Retrieves valid VPC peering authorizations that are pending for the AWS account \cr
\link[=gamelift_describe_vpc_peering_connections]{describe_vpc_peering_connections} \tab Retrieves information on VPC peering connections \cr
\link[=gamelift_get_game_session_log_url]{get_game_session_log_url} \tab Retrieves the location of stored game session logs for a specified game session \cr
\link[=gamelift_get_instance_access]{get_instance_access} \tab Requests remote access to a fleet instance \cr
\link[=gamelift_list_aliases]{list_aliases} \tab Retrieves all aliases for this AWS account \cr
\link[=gamelift_list_builds]{list_builds} \tab Retrieves build records for all builds associated with the AWS account in use \cr
\link[=gamelift_list_fleets]{list_fleets} \tab Retrieves a collection of fleet records for this AWS account \cr
\link[=gamelift_put_scaling_policy]{put_scaling_policy} \tab Creates or updates a scaling policy for a fleet \cr
\link[=gamelift_request_upload_credentials]{request_upload_credentials} \tab Retrieves a fresh set of credentials for use when uploading a new set of game build files to Amazon GameLift's Amazon S3 \cr
\link[=gamelift_resolve_alias]{resolve_alias} \tab Retrieves the fleet ID that a specified alias is currently pointing to \cr
\link[=gamelift_search_game_sessions]{search_game_sessions} \tab Retrieves all active game sessions that match a set of search criteria and sorts them in a specified order \cr
\link[=gamelift_start_fleet_actions]{start_fleet_actions} \tab Resumes activity on a fleet that was suspended with StopFleetActions \cr
\link[=gamelift_start_game_session_placement]{start_game_session_placement} \tab Places a request for a new game session in a queue (see CreateGameSessionQueue) \cr
\link[=gamelift_start_match_backfill]{start_match_backfill} \tab Finds new players to fill open slots in an existing game session \cr
\link[=gamelift_start_matchmaking]{start_matchmaking} \tab Uses FlexMatch to create a game match for a group of players based on custom matchmaking rules, and starts a new game for the matched players \cr
\link[=gamelift_stop_fleet_actions]{stop_fleet_actions} \tab Suspends activity on a fleet \cr
\link[=gamelift_stop_game_session_placement]{stop_game_session_placement} \tab Cancels a game session placement that is in PENDING status \cr
\link[=gamelift_stop_matchmaking]{stop_matchmaking} \tab Cancels a matchmaking ticket that is currently being processed \cr
\link[=gamelift_update_alias]{update_alias} \tab Updates properties for an alias \cr
\link[=gamelift_update_build]{update_build} \tab Updates metadata in a build record, including the build name and version \cr
\link[=gamelift_update_fleet_attributes]{update_fleet_attributes} \tab Updates fleet properties, including name and description, for a fleet \cr
\link[=gamelift_update_fleet_capacity]{update_fleet_capacity} \tab Updates capacity settings for a fleet \cr
\link[=gamelift_update_fleet_port_settings]{update_fleet_port_settings} \tab Updates port settings for a fleet \cr
\link[=gamelift_update_game_session]{update_game_session} \tab Updates game session properties \cr
\link[=gamelift_update_game_session_queue]{update_game_session_queue} \tab Updates settings for a game session queue, which determines how new game session requests in the queue are processed \cr
\link[=gamelift_update_matchmaking_configuration]{update_matchmaking_configuration} \tab Updates settings for a FlexMatch matchmaking configuration \cr
\link[=gamelift_update_runtime_configuration]{update_runtime_configuration} \tab Updates the current run-time configuration for the specified fleet, which tells Amazon GameLift how to launch server processes on instances in the fleet \cr
\link[=gamelift_validate_matchmaking_rule_set]{validate_matchmaking_rule_set} \tab Validates the syntax of a matchmaking rule or rule set
}
}

