% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_alias}
\alias{gamelift_delete_alias}
\title{Deletes an alias}
\usage{
gamelift_delete_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] A unique identifier of the alias that you want to delete. You can use
either the alias ID or ARN value.}
}
\description{
Deletes an alias. This action removes all record of the alias. Game
clients attempting to access a server process using the deleted alias
receive an error. To delete an alias, specify the alias ID to be
deleted.
}
\details{
\itemize{
\item CreateAlias
\item ListAliases
\item DescribeAlias
\item UpdateAlias
\item DeleteAlias
\item ResolveAlias
}
}
\section{Request syntax}{
\preformatted{svc$delete_alias(
  AliasId = "string"
)
}
}

\keyword{internal}
