% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_game_session_queue}
\alias{gamelift_delete_game_session_queue}
\title{Deletes a game session queue}
\usage{
gamelift_delete_game_session_queue(Name)
}
\arguments{
\item{Name}{[required] A descriptive label that is associated with game session queue. Queue
names must be unique within each Region. You can use either the queue ID
or ARN value.}
}
\description{
Deletes a game session queue. Once a queue is successfully deleted,
unfulfilled StartGameSessionPlacement requests that reference the queue
will fail. To delete a queue, specify the queue name.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/queues-intro.html}{Using Multi-Region Queues}

\strong{Related operations}
\itemize{
\item CreateGameSessionQueue
\item DescribeGameSessionQueues
\item UpdateGameSessionQueue
\item DeleteGameSessionQueue
}
}
\section{Request syntax}{
\preformatted{svc$delete_game_session_queue(
  Name = "string"
)
}
}

\keyword{internal}
