% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_matchmaking_rule_set}
\alias{gamelift_delete_matchmaking_rule_set}
\title{Deletes an existing matchmaking rule set}
\usage{
gamelift_delete_matchmaking_rule_set(Name)
}
\arguments{
\item{Name}{[required] A unique identifier for a matchmaking rule set to be deleted. (Note: The
rule set name is different from the optional "name" field in the rule
set body.) You can use either the rule set name or ARN value.}
}
\description{
Deletes an existing matchmaking rule set. To delete the rule set,
provide the rule set name. Rule sets cannot be deleted if they are
currently being used by a matchmaking configuration.
}
\details{
\strong{Learn more}
\itemize{
\item \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-rulesets.html}{Build a Rule Set}
}

\strong{Related operations}
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
\item DeleteMatchmakingRuleSet
}
}
\section{Request syntax}{
\preformatted{svc$delete_matchmaking_rule_set(
  Name = "string"
)
}
}

\keyword{internal}
