% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_script}
\alias{gamelift_delete_script}
\title{Deletes a Realtime script}
\usage{
gamelift_delete_script(ScriptId)
}
\arguments{
\item{ScriptId}{[required] A unique identifier for a Realtime script to delete. You can use either
the script ID or ARN value.}
}
\description{
Deletes a Realtime script. This action permanently deletes the script
record. If script files were uploaded, they are also deleted (files
stored in an S3 bucket are not deleted).
}
\details{
To delete a script, specify the script ID. Before deleting a script, be
sure to terminate all fleets that are deployed with the script being
deleted. Fleet instances periodically check for script updates, and if
the script record no longer exists, the instance will go into an error
state and be unable to host game sessions.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/realtime-intro.html}{Amazon GameLift Realtime Servers}

\strong{Related operations}
\itemize{
\item CreateScript
\item ListScripts
\item DescribeScript
\item UpdateScript
\item DeleteScript
}
}
\section{Request syntax}{
\preformatted{svc$delete_script(
  ScriptId = "string"
)
}
}

\keyword{internal}
