% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_vpc_peering_connection}
\alias{gamelift_delete_vpc_peering_connection}
\title{Removes a VPC peering connection}
\usage{
gamelift_delete_vpc_peering_connection(FleetId, VpcPeeringConnectionId)
}
\arguments{
\item{FleetId}{[required] A unique identifier for a fleet. This fleet specified must match the
fleet referenced in the VPC peering connection record. You can use
either the fleet ID or ARN value.}

\item{VpcPeeringConnectionId}{[required] A unique identifier for a VPC peering connection. This value is included
in the VpcPeeringConnection object, which can be retrieved by calling
DescribeVpcPeeringConnections.}
}
\description{
Removes a VPC peering connection. To delete the connection, you must
have a valid authorization for the VPC peering connection that you want
to delete. You can check for an authorization by calling
DescribeVpcPeeringAuthorizations or request a new one using
CreateVpcPeeringAuthorization.
}
\details{
Once a valid authorization exists, call this operation from the AWS
account that is used to manage the Amazon GameLift fleets. Identify the
connection to delete by the connection ID and fleet ID. If successful,
the connection is removed.
\itemize{
\item CreateVpcPeeringAuthorization
\item DescribeVpcPeeringAuthorizations
\item DeleteVpcPeeringAuthorization
\item CreateVpcPeeringConnection
\item DescribeVpcPeeringConnections
\item DeleteVpcPeeringConnection
}
}
\section{Request syntax}{
\preformatted{svc$delete_vpc_peering_connection(
  FleetId = "string",
  VpcPeeringConnectionId = "string"
)
}
}

\keyword{internal}
