% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_build}
\alias{gamelift_describe_build}
\title{Retrieves properties for a build}
\usage{
gamelift_describe_build(BuildId)
}
\arguments{
\item{BuildId}{[required] Unique identifier for a build to retrieve properties for.}
}
\description{
Retrieves properties for a build. To request a build record, specify a
build ID. If successful, an object containing the build properties is
returned.
}
\details{
\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/build-intro.html}{Working with Builds}

\strong{Related operations}
\itemize{
\item CreateBuild
\item ListBuilds
\item DescribeBuild
\item UpdateBuild
\item DeleteBuild
}
}
\section{Request syntax}{
\preformatted{gamelift$describe_build(
  BuildId = "string"
)
}
}

\keyword{internal}
