% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_port_settings}
\alias{gamelift_describe_fleet_port_settings}
\title{Retrieves a fleet's inbound connection permissions}
\usage{
gamelift_describe_fleet_port_settings(FleetId)
}
\arguments{
\item{FleetId}{[required] A unique identifier for a fleet to retrieve port settings for. You can
use either the fleet ID or ARN value.}
}
\description{
Retrieves a fleet's inbound connection permissions. Connection
permissions specify the range of IP addresses and port settings that
incoming traffic can use to access server processes in the fleet. Game
sessions that are running on instances in the fleet use connections that
fall in this range.

To get a fleet's inbound connection permissions, specify the fleet's
unique identifier. If successful, a collection of IpPermission objects
is returned for the requested fleet ID. If the requested fleet has been
deleted, the result set is empty.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\strong{Related operations}
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\itemize{
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
}
\item UpdateFleetAttributes
\item StartFleetActions or StopFleetActions
}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_port_settings(
  FleetId = "string"
)
}
}

\keyword{internal}
