% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_game_server_group}
\alias{gamelift_describe_game_server_group}
\title{This action is part of Amazon GameLift FleetIQ with game server groups,
which is in preview release and is subject to change}
\usage{
gamelift_describe_game_server_group(GameServerGroupName)
}
\arguments{
\item{GameServerGroupName}{[required] The unique identifier for the game server group being requested. Use
either the GameServerGroup name or ARN value.}
}
\description{
\strong{This action is part of Amazon GameLift FleetIQ with game server
groups, which is in preview release and is subject to change.}
}
\details{
Retrieves information on a game server group.

To get attributes for a game server group, provide a group name or ARN
value. If successful, a GameServerGroup object is returned.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item CreateGameServerGroup
\item ListGameServerGroups
\item DescribeGameServerGroup
\item UpdateGameServerGroup
\item DeleteGameServerGroup
\item ResumeGameServerGroup
\item SuspendGameServerGroup
}
}
\section{Request syntax}{
\preformatted{svc$describe_game_server_group(
  GameServerGroupName = "string"
)
}
}

\keyword{internal}
