% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_game_session_placement}
\alias{gamelift_describe_game_session_placement}
\title{Retrieves properties and current status of a game session placement
request}
\usage{
gamelift_describe_game_session_placement(PlacementId)
}
\arguments{
\item{PlacementId}{[required] Unique identifier for a game session placement to retrieve.}
}
\description{
Retrieves properties and current status of a game session placement
request. To get game session placement details, specify the placement
ID. If successful, a GameSessionPlacement object is returned.
}
\details{
\itemize{
\item CreateGameSession
\item DescribeGameSessions
\item DescribeGameSessionDetails
\item SearchGameSessions
\item UpdateGameSession
\item GetGameSessionLogUrl
\item Game session placements
\itemize{
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
}
\section{Request syntax}{
\preformatted{svc$describe_game_session_placement(
  PlacementId = "string"
)
}
}

\keyword{internal}
