% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_game_sessions}
\alias{gamelift_describe_game_sessions}
\title{Retrieves a set of one or more game sessions}
\usage{
gamelift_describe_game_sessions(FleetId, GameSessionId, AliasId,
  StatusFilter, Limit, NextToken)
}
\arguments{
\item{FleetId}{A unique identifier for a fleet to retrieve all game sessions for. You
can use either the fleet ID or ARN value.}

\item{GameSessionId}{A unique identifier for the game session to retrieve.}

\item{AliasId}{A unique identifier for an alias associated with the fleet to retrieve
all game sessions for. You can use either the alias ID or ARN value.}

\item{StatusFilter}{Game session status to filter results on. Possible game session statuses
include \code{ACTIVE}, \code{TERMINATED}, \code{ACTIVATING}, and \code{TERMINATING} (the
last two are transitory).}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{Token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves a set of one or more game sessions. Request a specific game
session or request all game sessions on a fleet. Alternatively, use
SearchGameSessions to request a set of active game sessions that are
filtered by certain criteria. To retrieve protection policy settings for
game sessions, use DescribeGameSessionDetails.
}
\details{
To get game sessions, specify one of the following: game session ID,
fleet ID, or alias ID. You can filter this request by game session
status. Use the pagination parameters to retrieve results as a set of
sequential pages. If successful, a GameSession object is returned for
each game session matching the request.

\emph{Available in Amazon GameLift Local.}
\itemize{
\item CreateGameSession
\item DescribeGameSessions
\item DescribeGameSessionDetails
\item SearchGameSessions
\item UpdateGameSession
\item GetGameSessionLogUrl
\item Game session placements
\itemize{
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
}
\section{Request syntax}{
\preformatted{svc$describe_game_sessions(
  FleetId = "string",
  GameSessionId = "string",
  AliasId = "string",
  StatusFilter = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
