% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_matchmaking}
\alias{gamelift_describe_matchmaking}
\title{Retrieves one or more matchmaking tickets}
\usage{
gamelift_describe_matchmaking(TicketIds)
}
\arguments{
\item{TicketIds}{[required] Unique identifier for a matchmaking ticket. You can include up to 10 ID
values.}
}
\description{
Retrieves one or more matchmaking tickets. Use this operation to
retrieve ticket information, including status and\--once a successful
match is made\--acquire connection information for the resulting new
game session.
}
\details{
You can use this operation to track the progress of matchmaking requests
(through polling) as an alternative to using event notifications. See
more details on tracking matchmaking requests through polling or
notifications in StartMatchmaking.

To request matchmaking tickets, provide a list of up to 10 ticket IDs.
If the request is successful, a ticket object is returned for each
requested ID that currently exists.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/match-client.html}{Add FlexMatch to a Game Client}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguidematch-notification.html}{Set Up FlexMatch Event Notification}

\strong{Related operations}
\itemize{
\item StartMatchmaking
\item DescribeMatchmaking
\item StopMatchmaking
\item AcceptMatch
\item StartMatchBackfill
}
}
\section{Request syntax}{
\preformatted{svc$describe_matchmaking(
  TicketIds = list(
    "string"
  )
)
}
}

\keyword{internal}
