% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_matchmaking_configurations}
\alias{gamelift_describe_matchmaking_configurations}
\title{Retrieves the details of FlexMatch matchmaking configurations}
\usage{
gamelift_describe_matchmaking_configurations(Names, RuleSetName, Limit,
  NextToken)
}
\arguments{
\item{Names}{A unique identifier for a matchmaking configuration(s) to retrieve. You
can use either the configuration name or ARN value. To request all
existing configurations, leave this parameter empty.}

\item{RuleSetName}{A unique identifier for a matchmaking rule set. You can use either the
rule set name or ARN value. Use this parameter to retrieve all
matchmaking configurations that use this rule set.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages. This parameter
is limited to 10.}

\item{NextToken}{A token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves the details of FlexMatch matchmaking configurations.

This operation offers the following options: (1) retrieve all
matchmaking configurations, (2) retrieve configurations for a specified
list, or (3) retrieve all configurations that use a specified rule set
name. When requesting multiple items, use the pagination parameters to
retrieve results as a set of sequential pages.

If successful, a configuration is returned for each requested name. When
specifying a list of names, only configurations that currently exist are
returned.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/matchmaker-build.html}{Setting Up FlexMatch Matchmakers}

\strong{Related operations}
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
\item DeleteMatchmakingRuleSet
}
}
\section{Request syntax}{
\preformatted{svc$describe_matchmaking_configurations(
  Names = list(
    "string"
  ),
  RuleSetName = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
