% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_runtime_configuration}
\alias{gamelift_describe_runtime_configuration}
\title{Retrieves a fleet's runtime configuration settings}
\usage{
gamelift_describe_runtime_configuration(FleetId)
}
\arguments{
\item{FleetId}{[required] A unique identifier for a fleet to get the runtime configuration for.
You can use either the fleet ID or ARN value.}
}
\description{
Retrieves a fleet\'s runtime configuration settings. The runtime
configuration tells Amazon GameLift which server processes to run (and
how) on each instance in the fleet.
}
\details{
To get a runtime configuration, specify the fleet\'s unique identifier.
If successful, a RuntimeConfiguration object is returned for the
requested fleet. If the requested fleet has been deleted, the result set
is empty.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-multiprocess.html}{Running Multiple Processes on a Fleet}

\strong{Related operations}
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\itemize{
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
}
\item UpdateFleetAttributes
\item StartFleetActions or StopFleetActions
}
}
\section{Request syntax}{
\preformatted{svc$describe_runtime_configuration(
  FleetId = "string"
)
}
}

\keyword{internal}
