% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_vpc_peering_authorizations}
\alias{gamelift_describe_vpc_peering_authorizations}
\title{Retrieves valid VPC peering authorizations that are pending for the AWS
account}
\usage{
gamelift_describe_vpc_peering_authorizations()
}
\description{
Retrieves valid VPC peering authorizations that are pending for the AWS
account. This operation returns all VPC peering authorizations and
requests for peering. This includes those initiated and received by this
account.
\itemize{
\item \code{\link[=gamelift_create_vpc_peering_authorization]{create_vpc_peering_authorization}}
\item \code{\link[=gamelift_describe_vpc_peering_authorizations]{describe_vpc_peering_authorizations}}
\item \code{\link[=gamelift_delete_vpc_peering_authorization]{delete_vpc_peering_authorization}}
\item \code{\link[=gamelift_create_vpc_peering_connection]{create_vpc_peering_connection}}
\item \code{\link[=gamelift_describe_vpc_peering_connections]{describe_vpc_peering_connections}}
\item \code{\link[=gamelift_delete_vpc_peering_connection]{delete_vpc_peering_connection}}
}
}
\section{Request syntax}{
\preformatted{svc$describe_vpc_peering_authorizations()
}
}

\keyword{internal}
