% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_list_aliases}
\alias{gamelift_list_aliases}
\title{Retrieves all aliases for this AWS account}
\usage{
gamelift_list_aliases(RoutingStrategyType, Name, Limit, NextToken)
}
\arguments{
\item{RoutingStrategyType}{Type of routing to filter results on. Use this parameter to retrieve
only aliases of a certain type. To retrieve all aliases, leave this
parameter empty.

Possible routing types include the following:
\itemize{
\item \strong{SIMPLE} \\-- The alias resolves to one specific fleet. Use this
type when routing to active fleets.
\item \strong{TERMINAL} \\-- The alias does not resolve to a fleet but instead
can be used to display a message to the user. A terminal alias
throws a TerminalRoutingStrategyException with the RoutingStrategy
message embedded.
}}

\item{Name}{Descriptive label that is associated with an alias. Alias names do not
need to be unique.}

\item{Limit}{Maximum number of results to return. Use this parameter with \code{NextToken}
to get results as a set of sequential pages.}

\item{NextToken}{Token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this action. To
start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves all aliases for this AWS account. You can filter the result
set by alias name and/or routing strategy type. Use the pagination
parameters to retrieve results in sequential pages.
}
\details{
Returned aliases are not listed in any particular order.
\itemize{
\item CreateAlias
\item ListAliases
\item DescribeAlias
\item UpdateAlias
\item DeleteAlias
\item ResolveAlias
}
}
\section{Request syntax}{
\preformatted{svc$list_aliases(
  RoutingStrategyType = "SIMPLE"|"TERMINAL",
  Name = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
