% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_resolve_alias}
\alias{gamelift_resolve_alias}
\title{Retrieves the fleet ID that an alias is currently pointing to}
\usage{
gamelift_resolve_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] The unique identifier of the alias that you want to retrieve a fleet ID
for. You can use either the alias ID or ARN value.}
}
\description{
Retrieves the fleet ID that an alias is currently pointing to.
\itemize{
\item CreateAlias
\item ListAliases
\item DescribeAlias
\item UpdateAlias
\item DeleteAlias
\item ResolveAlias
}
}
\section{Request syntax}{
\preformatted{svc$resolve_alias(
  AliasId = "string"
)
}
}

\keyword{internal}
