% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_start_fleet_actions}
\alias{gamelift_start_fleet_actions}
\title{Resumes activity on a fleet that was suspended with StopFleetActions}
\usage{
gamelift_start_fleet_actions(FleetId, Actions)
}
\arguments{
\item{FleetId}{[required] A unique identifier for a fleet to start actions on. You can use either
the fleet ID or ARN value.}

\item{Actions}{[required] List of actions to restart on the fleet.}
}
\description{
Resumes activity on a fleet that was suspended with StopFleetActions.
Currently, this operation is used to restart a fleet's auto-scaling
activity.
}
\details{
To start fleet actions, specify the fleet ID and the type of actions to
restart. When auto-scaling fleet actions are restarted, Amazon GameLift
once again initiates scaling events as triggered by the fleet's scaling
policies. If actions on the fleet were never stopped, this operation
will have no effect. You can view a fleet's stopped actions using
DescribeFleetAttributes.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\strong{Related operations}
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item DescribeFleetAttributes
\item UpdateFleetAttributes
\item StartFleetActions or StopFleetActions
}
}
\section{Request syntax}{
\preformatted{svc$start_fleet_actions(
  FleetId = "string",
  Actions = list(
    "AUTO_SCALING"
  )
)
}
}

\keyword{internal}
