% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_stop_game_session_placement}
\alias{gamelift_stop_game_session_placement}
\title{Cancels a game session placement that is in PENDING status}
\usage{
gamelift_stop_game_session_placement(PlacementId)
}
\arguments{
\item{PlacementId}{[required] A unique identifier for a game session placement to cancel.}
}
\description{
Cancels a game session placement that is in \code{PENDING} status. To stop a
placement, provide the placement ID values. If successful, the placement
is moved to \code{CANCELLED} status.
\itemize{
\item \code{\link[=gamelift_create_game_session]{create_game_session}}
\item \code{\link[=gamelift_describe_game_sessions]{describe_game_sessions}}
\item \code{\link[=gamelift_describe_game_session_details]{describe_game_session_details}}
\item \code{\link[=gamelift_search_game_sessions]{search_game_sessions}}
\item \code{\link[=gamelift_update_game_session]{update_game_session}}
\item \code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}}
\item Game session placements
\itemize{
\item \code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
\item \code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
\item \code{\link[=gamelift_stop_game_session_placement]{stop_game_session_placement}}
}
}
}
\section{Request syntax}{
\preformatted{svc$stop_game_session_placement(
  PlacementId = "string"
)
}
}

\keyword{internal}
