% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_alias}
\alias{gamelift_update_alias}
\title{Updates properties for an alias}
\usage{
gamelift_update_alias(AliasId, Name, Description, RoutingStrategy)
}
\arguments{
\item{AliasId}{[required] A unique identifier for the alias that you want to update. You can use
either the alias ID or ARN value.}

\item{Name}{A descriptive label that is associated with an alias. Alias names do not
need to be unique.}

\item{Description}{A human-readable description of the alias.}

\item{RoutingStrategy}{The routing configuration, including routing type and fleet target, for
the alias.}
}
\description{
Updates properties for an alias. To update properties, specify the alias
ID to be updated and provide the information to be changed. To reassign
an alias to another fleet, provide an updated routing strategy. If
successful, the updated alias record is returned.
\itemize{
\item \code{\link[=gamelift_create_alias]{create_alias}}
\item \code{\link[=gamelift_list_aliases]{list_aliases}}
\item \code{\link[=gamelift_describe_alias]{describe_alias}}
\item \code{\link[=gamelift_update_alias]{update_alias}}
\item \code{\link[=gamelift_delete_alias]{delete_alias}}
\item \code{\link[=gamelift_resolve_alias]{resolve_alias}}
}
}
\section{Request syntax}{
\preformatted{svc$update_alias(
  AliasId = "string",
  Name = "string",
  Description = "string",
  RoutingStrategy = list(
    Type = "SIMPLE"|"TERMINAL",
    FleetId = "string",
    Message = "string"
  )
)
}
}

\keyword{internal}
