% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_fleet_capacity}
\alias{gamelift_update_fleet_capacity}
\title{Updates capacity settings for a fleet}
\usage{
gamelift_update_fleet_capacity(FleetId, DesiredInstances, MinSize,
  MaxSize)
}
\arguments{
\item{FleetId}{[required] Unique identifier for a fleet to update capacity for.}

\item{DesiredInstances}{Number of EC2 instances you want this fleet to host.}

\item{MinSize}{Minimum value allowed for the fleet's instance count. Default if not
set is 0.}

\item{MaxSize}{Maximum value allowed for the fleet's instance count. Default if not
set is 1.}
}
\description{
Updates capacity settings for a fleet. Use this action to specify the
number of EC2 instances (hosts) that you want this fleet to contain.
Before calling this action, you may want to call
DescribeEC2InstanceLimits to get the maximum capacity based on the
fleet's EC2 instance type.
}
\details{
Specify minimum and maximum number of instances. Amazon GameLift will
not change fleet capacity to values fall outside of this range. This is
particularly important when using auto-scaling (see PutScalingPolicy) to
allow capacity to adjust based on player demand while imposing limits on
automatic adjustments.

To update fleet capacity, specify the fleet ID and the number of
instances you want the fleet to host. If successful, Amazon GameLift
starts or terminates instances so that the fleet's active instance
count matches the desired instance count. You can view a fleet's
current capacity information by calling DescribeFleetCapacity. If the
desired instance count is higher than the instance type's limit, the
"Limit Exceeded" exception occurs.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Working with Fleets}.

\strong{Related operations}
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
\item Update fleets:
\item UpdateFleetAttributes
\item UpdateFleetCapacity
\item UpdateFleetPortSettings
\item UpdateRuntimeConfiguration
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\section{Request syntax}{
\preformatted{gamelift$update_fleet_capacity(
  FleetId = "string",
  DesiredInstances = 123,
  MinSize = 123,
  MaxSize = 123
)
}
}

\keyword{internal}
