% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_game_server_group}
\alias{gamelift_update_game_server_group}
\title{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups}
\usage{
gamelift_update_game_server_group(GameServerGroupName, RoleArn,
  InstanceDefinitions, GameServerProtectionPolicy, BalancingStrategy)
}
\arguments{
\item{GameServerGroupName}{[required] A unique identifier for the game server group. Use either the
GameServerGroup name or ARN value.}

\item{RoleArn}{The Amazon Resource Name
(\href{https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-arn-format.html}{ARN})
for an IAM role that allows Amazon GameLift to access your EC2 Auto
Scaling groups.}

\item{InstanceDefinitions}{An updated list of EC2 instance types to use in the Auto Scaling group.
The instance definitions must specify at least two different instance
types that are supported by GameLift FleetIQ. This updated list replaces
the entire current list of instance definitions for the game server
group. For more information on instance types, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html}{EC2 Instance Types}
in the \emph{Amazon EC2 User Guide}. You can optionally specify capacity
weighting for each instance type. If no weight value is specified for an
instance type, it is set to the default value "1". For more information
about capacity weighting, see \href{https://docs.aws.amazon.com/autoscaling/ec2/userguide/asg-instance-weighting.html}{Instance Weighting for Amazon EC2 Auto Scaling}
in the Amazon EC2 Auto Scaling User Guide.}

\item{GameServerProtectionPolicy}{A flag that indicates whether instances in the game server group are
protected from early termination. Unprotected instances that have active
game servers running might be terminated during a scale-down event,
causing players to be dropped from the game. Protected instances cannot
be terminated while there are active game servers running except in the
event of a forced game server group deletion (see ). An exception to
this is with Spot Instances, which can be terminated by AWS regardless
of protection status. This property is set to \code{NO_PROTECTION} by
default.}

\item{BalancingStrategy}{Indicates how GameLift FleetIQ balances the use of Spot Instances and
On-Demand Instances in the game server group. Method options include the
following:
\itemize{
\item \code{SPOT_ONLY} - Only Spot Instances are used in the game server group.
If Spot Instances are unavailable or not viable for game hosting,
the game server group provides no hosting capacity until Spot
Instances can again be used. Until then, no new instances are
started, and the existing nonviable Spot Instances are terminated
(after current gameplay ends) and are not replaced.
\item \code{SPOT_PREFERRED} - (default value) Spot Instances are used whenever
available in the game server group. If Spot Instances are
unavailable, the game server group continues to provide hosting
capacity by falling back to On-Demand Instances. Existing nonviable
Spot Instances are terminated (after current gameplay ends) and are
replaced with new On-Demand Instances.
\item \code{ON_DEMAND_ONLY} - Only On-Demand Instances are used in the game
server group. No Spot Instances are used, even when available, while
this balancing strategy is in force.
}}
}
\description{
\strong{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups.}
}
\details{
Updates GameLift FleetIQ-specific properties for a game server group.
Many Auto Scaling group properties are updated on the Auto Scaling group
directly, including the launch template, Auto Scaling policies, and
maximum/minimum/desired instance counts.

To update the game server group, specify the game server group ID and
provide the updated values. Before applying the updates, the new values
are validated to ensure that GameLift FleetIQ can continue to perform
instance balancing activity. If successful, a GameServerGroup object is
returned.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item CreateGameServerGroup
\item ListGameServerGroups
\item DescribeGameServerGroup
\item UpdateGameServerGroup
\item DeleteGameServerGroup
\item ResumeGameServerGroup
\item SuspendGameServerGroup
\item DescribeGameServerInstances
}
}
\section{Request syntax}{
\preformatted{svc$update_game_server_group(
  GameServerGroupName = "string",
  RoleArn = "string",
  InstanceDefinitions = list(
    list(
      InstanceType = "c4.large"|"c4.xlarge"|"c4.2xlarge"|"c4.4xlarge"|"c4.8xlarge"|"c5.large"|"c5.xlarge"|"c5.2xlarge"|"c5.4xlarge"|"c5.9xlarge"|"c5.12xlarge"|"c5.18xlarge"|"c5.24xlarge"|"r4.large"|"r4.xlarge"|"r4.2xlarge"|"r4.4xlarge"|"r4.8xlarge"|"r4.16xlarge"|"r5.large"|"r5.xlarge"|"r5.2xlarge"|"r5.4xlarge"|"r5.8xlarge"|"r5.12xlarge"|"r5.16xlarge"|"r5.24xlarge"|"m4.large"|"m4.xlarge"|"m4.2xlarge"|"m4.4xlarge"|"m4.10xlarge"|"m5.large"|"m5.xlarge"|"m5.2xlarge"|"m5.4xlarge"|"m5.8xlarge"|"m5.12xlarge"|"m5.16xlarge"|"m5.24xlarge",
      WeightedCapacity = "string"
    )
  ),
  GameServerProtectionPolicy = "NO_PROTECTION"|"FULL_PROTECTION",
  BalancingStrategy = "SPOT_ONLY"|"SPOT_PREFERRED"|"ON_DEMAND_ONLY"
)
}
}

\keyword{internal}
