% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_game_session}
\alias{gamelift_update_game_session}
\title{Updates game session properties}
\usage{
gamelift_update_game_session(GameSessionId, MaximumPlayerSessionCount,
  Name, PlayerSessionCreationPolicy, ProtectionPolicy)
}
\arguments{
\item{GameSessionId}{[required] Unique identifier for the game session to update.}

\item{MaximumPlayerSessionCount}{Maximum number of players that can be connected simultaneously to the
game session.}

\item{Name}{Descriptive label that is associated with a game session. Session names
do not need to be unique.}

\item{PlayerSessionCreationPolicy}{Policy determining whether or not the game session accepts new players.}

\item{ProtectionPolicy}{Game session protection policy to apply to this game session only.
\itemize{
\item \strong{NoProtection} \-- The game session can be terminated during a
scale-down event.
\item \strong{FullProtection} \-- If the game session is in an \code{ACTIVE} status,
it cannot be terminated during a scale-down event.
}}
}
\description{
Updates game session properties. This includes the session name, maximum
player count, protection policy, which controls whether or not an active
game session can be terminated during a scale-down event, and the player
session creation policy, which controls whether or not new players can
join the session. To update a game session, specify the game session ID
and the values you want to change. If successful, an updated GameSession
object is returned.
}
\details{
\itemize{
\item CreateGameSession
\item DescribeGameSessions
\item DescribeGameSessionDetails
\item SearchGameSessions
\item UpdateGameSession
\item GetGameSessionLogUrl
\item Game session placements
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
\section{Request syntax}{
\preformatted{svc$update_game_session(
  GameSessionId = "string",
  MaximumPlayerSessionCount = 123,
  Name = "string",
  PlayerSessionCreationPolicy = "ACCEPT_ALL"|"DENY_ALL",
  ProtectionPolicy = "NoProtection"|"FullProtection"
)
}
}

\keyword{internal}
