% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_game_session_queue}
\alias{gamelift_update_game_session_queue}
\title{Updates settings for a game session queue, which determines how new game
session requests in the queue are processed}
\usage{
gamelift_update_game_session_queue(Name, TimeoutInSeconds,
  PlayerLatencyPolicies, Destinations)
}
\arguments{
\item{Name}{[required] Descriptive label that is associated with game session queue. Queue
names must be unique within each region.}

\item{TimeoutInSeconds}{Maximum time, in seconds, that a new game session placement request
remains in the queue. When a request exceeds this time, the game session
placement changes to a \code{TIMED_OUT} status.}

\item{PlayerLatencyPolicies}{Collection of latency policies to apply when processing game sessions
placement requests with player latency information. Multiple policies
are evaluated in order of the maximum latency value, starting with the
lowest latency values. With just one policy, it is enforced at the start
of the game session placement for the duration period. With multiple
policies, each policy is enforced consecutively for its duration period.
For example, a queue might enforce a 60-second policy followed by a
120-second policy, and then no policy for the remainder of the
placement. When updating policies, provide a complete collection of
policies.}

\item{Destinations}{List of fleets that can be used to fulfill game session placement
requests in the queue. Fleets are identified by either a fleet ARN or a
fleet alias ARN. Destinations are listed in default preference order.
When updating this list, provide a complete list of destinations.}
}
\description{
Updates settings for a game session queue, which determines how new game
session requests in the queue are processed. To update settings, specify
the queue name to be updated and provide the new settings. When updating
destinations, provide a complete list of destinations.
}
\details{
\itemize{
\item CreateGameSessionQueue
\item DescribeGameSessionQueues
\item UpdateGameSessionQueue
\item DeleteGameSessionQueue
}
}
\section{Request syntax}{
\preformatted{svc$update_game_session_queue(
  Name = "string",
  TimeoutInSeconds = 123,
  PlayerLatencyPolicies = list(
    list(
      MaximumIndividualPlayerLatencyMilliseconds = 123,
      PolicyDurationSeconds = 123
    )
  ),
  Destinations = list(
    list(
      DestinationArn = "string"
    )
  )
)
}
}

\keyword{internal}
