% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_matchmaking_configuration}
\alias{gamelift_update_matchmaking_configuration}
\title{Updates settings for a FlexMatch matchmaking configuration}
\usage{
gamelift_update_matchmaking_configuration(Name, Description,
  GameSessionQueueArns, RequestTimeoutSeconds, AcceptanceTimeoutSeconds,
  AcceptanceRequired, RuleSetName, NotificationTarget,
  AdditionalPlayerCount, CustomEventData, GameProperties, GameSessionData,
  BackfillMode, FlexMatchMode)
}
\arguments{
\item{Name}{[required] A unique identifier for a matchmaking configuration to update. You can
use either the configuration name or ARN value.}

\item{Description}{A descriptive label that is associated with matchmaking configuration.}

\item{GameSessionQueueArns}{Amazon Resource Name
(\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN})
that is assigned to a GameLift game session queue resource and uniquely
identifies it. ARNs are unique across all Regions. Queues can be located
in any Region. Queues are used to start new GameLift-hosted game
sessions for matches that are created with this matchmaking
configuration. If \code{FlexMatchMode} is set to \code{STANDALONE}, do not set
this parameter.}

\item{RequestTimeoutSeconds}{The maximum duration, in seconds, that a matchmaking ticket can remain
in process before timing out. Requests that fail due to timing out can
be resubmitted as needed.}

\item{AcceptanceTimeoutSeconds}{The length of time (in seconds) to wait for players to accept a proposed
match, if acceptance is required. If any player rejects the match or
fails to accept before the timeout, the tickets are returned to the
ticket pool and continue to be evaluated for an acceptable match.}

\item{AcceptanceRequired}{A flag that indicates whether a match that was created with this
configuration must be accepted by the matched players. To require
acceptance, set to TRUE. With this option enabled, matchmaking tickets
use the status \code{REQUIRES_ACCEPTANCE} to indicate when a completed
potential match is waiting for player acceptance.}

\item{RuleSetName}{A unique identifier for a matchmaking rule set to use with this
configuration. You can use either the rule set name or ARN value. A
matchmaking configuration can only use rule sets that are defined in the
same Region.}

\item{NotificationTarget}{An SNS topic ARN that is set up to receive matchmaking notifications.
See \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-notification.html}{Setting up Notifications for Matchmaking}
for more information.}

\item{AdditionalPlayerCount}{The number of player slots in a match to keep open for future players.
For example, assume that the configuration's rule set specifies a match
for a single 12-person team. If the additional player count is set to 2,
only 10 players are initially selected for the match. This parameter is
not used if \code{FlexMatchMode} is set to \code{STANDALONE}.}

\item{CustomEventData}{Information to add to all events related to the matchmaking
configuration.}

\item{GameProperties}{A set of custom properties for a game session, formatted as key-value
pairs. These properties are passed to a game server process in the
GameSession object with a request to start a new game session (see
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).
This information is added to the new GameSession object that is created
for a successful match. This parameter is not used if \code{FlexMatchMode} is
set to \code{STANDALONE}.}

\item{GameSessionData}{A set of custom game session properties, formatted as a single string
value. This data is passed to a game server process in the GameSession
object with a request to start a new game session (see \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).
This information is added to the new GameSession object that is created
for a successful match. This parameter is not used if \code{FlexMatchMode} is
set to \code{STANDALONE}.}

\item{BackfillMode}{The method that is used to backfill game sessions created with this
matchmaking configuration. Specify MANUAL when your game manages
backfill requests manually or does not use the match backfill feature.
Specify AUTOMATIC to have GameLift create a StartMatchBackfill request
whenever a game session has one or more open slots. Learn more about
manual and automatic backfill in \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-backfill.html}{Backfill Existing Games with FlexMatch}.
Automatic backfill is not available when \code{FlexMatchMode} is set to
\code{STANDALONE}.}

\item{FlexMatchMode}{Indicates whether this matchmaking configuration is being used with
GameLift hosting or as a standalone matchmaking solution.
\itemize{
\item \strong{STANDALONE} - FlexMatch forms matches and returns match
information, including players and team assignments, in a
\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-events.html#match-events-matchmakingsucceeded}{MatchmakingSucceeded}
event.
\item \strong{WITH\\_QUEUE} - FlexMatch forms matches and uses the specified
GameLift queue to start a game session for the match.
}}
}
\description{
Updates settings for a FlexMatch matchmaking configuration. These
changes affect all matches and game sessions that are created after the
update. To update settings, specify the configuration name to be updated
and provide the new settings.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-configuration.html}{Design a FlexMatch Matchmaker}

\strong{Related operations}
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
\item DeleteMatchmakingRuleSet
}
}
\section{Request syntax}{
\preformatted{svc$update_matchmaking_configuration(
  Name = "string",
  Description = "string",
  GameSessionQueueArns = list(
    "string"
  ),
  RequestTimeoutSeconds = 123,
  AcceptanceTimeoutSeconds = 123,
  AcceptanceRequired = TRUE|FALSE,
  RuleSetName = "string",
  NotificationTarget = "string",
  AdditionalPlayerCount = 123,
  CustomEventData = "string",
  GameProperties = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  GameSessionData = "string",
  BackfillMode = "AUTOMATIC"|"MANUAL",
  FlexMatchMode = "STANDALONE"|"WITH_QUEUE"
)
}
}

\keyword{internal}
