% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_accelerator_attributes}
\alias{globalaccelerator_describe_accelerator_attributes}
\title{Describe the attributes of an accelerator}
\usage{
globalaccelerator_describe_accelerator_attributes(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator with the attributes
that you want to describe.}
}
\description{
Describe the attributes of an accelerator. To see an AWS CLI example of
describing the attributes of an accelerator, scroll down to \strong{Example}.
}
\section{Request syntax}{
\preformatted{svc$describe_accelerator_attributes(
  AcceleratorArn = "string"
)
}
}

\keyword{internal}
