% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_endpoint_group}
\alias{globalaccelerator_describe_endpoint_group}
\title{Describe an endpoint group}
\usage{
globalaccelerator_describe_endpoint_group(EndpointGroupArn)
}
\arguments{
\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group to describe.}
}
\description{
Describe an endpoint group.
}
\section{Request syntax}{
\preformatted{svc$describe_endpoint_group(
  EndpointGroupArn = "string"
)
}
}

\keyword{internal}
