% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_listeners}
\alias{globalaccelerator_list_listeners}
\title{List the listeners for an accelerator}
\usage{
globalaccelerator_list_listeners(AcceleratorArn, MaxResults, NextToken)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator for which you want to
list listener objects.}

\item{MaxResults}{The number of listener objects that you want to return with this call.
The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a
previous call.}
}
\description{
List the listeners for an accelerator. To see an AWS CLI example of
listing the listeners for an accelerator, scroll down to \strong{Example}.
}
\section{Request syntax}{
\preformatted{svc$list_listeners(
  AcceleratorArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
