% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_jobs}
\alias{glue_batch_get_jobs}
\title{Returns a list of resource metadata for a given list of job names}
\usage{
glue_batch_get_jobs(JobNames)
}
\arguments{
\item{JobNames}{[required] A list of job names, which may be the names returned from the \code{ListJobs}
operation.}
}
\description{
Returns a list of resource metadata for a given list of job names. After
calling the \code{ListJobs} operation, you can call this operation to access
the data to which you have been granted permissions. This operation
supports all IAM permissions, including permission conditions that uses
tags.
}
\section{Request syntax}{
\preformatted{glue$batch_get_jobs(
  JobNames = list(
    "string"
  )
)
}
}

\keyword{internal}
