% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_stop_job_run}
\alias{glue_batch_stop_job_run}
\title{Stops one or more job runs for a specified job definition}
\usage{
glue_batch_stop_job_run(JobName, JobRunIds)
}
\arguments{
\item{JobName}{[required] The name of the job definition for which to stop job runs.}

\item{JobRunIds}{[required] A list of the \code{JobRunIds} that should be stopped for that job
definition.}
}
\description{
Stops one or more job runs for a specified job definition.
}
\section{Request syntax}{
\preformatted{svc$batch_stop_job_run(
  JobName = "string",
  JobRunIds = list(
    "string"
  )
)
}
}

\keyword{internal}
