% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_classifier}
\alias{glue_create_classifier}
\title{Creates a classifier in the user's account}
\usage{
glue_create_classifier(GrokClassifier, XMLClassifier, JsonClassifier)
}
\arguments{
\item{GrokClassifier}{A \code{GrokClassifier} object specifying the classifier to create.}

\item{XMLClassifier}{An \code{XMLClassifier} object specifying the classifier to create.}

\item{JsonClassifier}{A \code{JsonClassifier} object specifying the classifier to create.}
}
\description{
Creates a classifier in the user's account. This may be a
\code{GrokClassifier}, an \code{XMLClassifier}, or abbrev \code{JsonClassifier},
depending on which field of the request is present.
}
\section{Request syntax}{
\preformatted{glue$create_classifier(
  GrokClassifier = list(
    Classification = "string",
    Name = "string",
    GrokPattern = "string",
    CustomPatterns = "string"
  ),
  XMLClassifier = list(
    Classification = "string",
    Name = "string",
    RowTag = "string"
  ),
  JsonClassifier = list(
    Name = "string",
    JsonPath = "string"
  )
)
}
}

\keyword{internal}
