% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_classifier}
\alias{glue_delete_classifier}
\title{Removes a classifier from the Data Catalog}
\usage{
glue_delete_classifier(Name)
}
\arguments{
\item{Name}{[required] Name of the classifier to remove.}
}
\description{
Removes a classifier from the Data Catalog.
}
\section{Request syntax}{
\preformatted{glue$delete_classifier(
  Name = "string"
)
}
}

\keyword{internal}
