% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_column_statistics_for_table}
\alias{glue_delete_column_statistics_for_table}
\title{Retrieves table statistics of columns}
\usage{
glue_delete_column_statistics_for_table(CatalogId, DatabaseName,
  TableName, ColumnName)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partitions in question reside. If
none is supplied, the AWS account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the partitions reside.}

\item{TableName}{[required] The name of the partitions' table.}

\item{ColumnName}{[required] The name of the column.}
}
\description{
Retrieves table statistics of columns.
}
\details{
The Identity and Access Management (IAM) permission required for this
operation is \code{DeleteTable}.
}
\section{Request syntax}{
\preformatted{svc$delete_column_statistics_for_table(
  CatalogId = "string",
  DatabaseName = "string",
  TableName = "string",
  ColumnName = "string"
)
}
}

\keyword{internal}
