% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_partition_index}
\alias{glue_delete_partition_index}
\title{Deletes a specified partition index from an existing table}
\usage{
glue_delete_partition_index(CatalogId, DatabaseName, TableName,
  IndexName)
}
\arguments{
\item{CatalogId}{The catalog ID where the table resides.}

\item{DatabaseName}{[required] Specifies the name of a database from which you want to delete a
partition index.}

\item{TableName}{[required] Specifies the name of a table from which you want to delete a partition
index.}

\item{IndexName}{[required] The name of the partition index to be deleted.}
}
\description{
Deletes a specified partition index from an existing table.
}
\section{Request syntax}{
\preformatted{svc$delete_partition_index(
  CatalogId = "string",
  DatabaseName = "string",
  TableName = "string",
  IndexName = "string"
)
}
}

\keyword{internal}
