% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_classifier}
\alias{glue_get_classifier}
\title{Retrieve a classifier by name}
\usage{
glue_get_classifier(Name)
}
\arguments{
\item{Name}{[required] Name of the classifier to retrieve.}
}
\description{
Retrieve a classifier by name.
}
\section{Request syntax}{
\preformatted{glue$get_classifier(
  Name = "string"
)
}
}

\keyword{internal}
