% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_connections}
\alias{glue_get_connections}
\title{Retrieves a list of connection definitions from the Data Catalog}
\usage{
glue_get_connections(CatalogId, Filter, HidePassword, NextToken,
  MaxResults)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which the connections reside. If none is
supplied, the AWS account ID is used by default.}

\item{Filter}{A filter that controls which connections will be returned.}

\item{HidePassword}{Allow you to retrieve the connection metadata without displaying the
password. For instance, the AWS Glue console uses this flag to retrieve
connections, since the console does not display passwords. Set this
parameter where the caller may not have permission to use the KMS key to
decrypt the password, but does have permission to access the rest of the
connection metadata (that is, the other connection properties).}

\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum number of connections to return in one response.}
}
\description{
Retrieves a list of connection definitions from the Data Catalog.
}
\section{Request syntax}{
\preformatted{glue$get_connections(
  CatalogId = "string",
  Filter = list(
    MatchCriteria = list(
      "string"
    ),
    ConnectionType = "JDBC"|"SFTP"
  ),
  HidePassword = TRUE|FALSE,
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
