% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_databases}
\alias{glue_get_databases}
\title{Retrieves all Databases defined in a given Data Catalog}
\usage{
glue_get_databases(CatalogId, NextToken, MaxResults)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog from which to retrieve \code{Databases}. If none
is supplied, the AWS account ID is used by default.}

\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum number of databases to return in one response.}
}
\description{
Retrieves all Databases defined in a given Data Catalog.
}
\section{Request syntax}{
\preformatted{glue$get_databases(
  CatalogId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
