% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_job_bookmark}
\alias{glue_get_job_bookmark}
\title{Returns information on a job bookmark entry}
\usage{
glue_get_job_bookmark(JobName, RunId)
}
\arguments{
\item{JobName}{[required] The name of the job in question.}

\item{RunId}{The unique run identifier associated with this job run.}
}
\description{
Returns information on a job bookmark entry.
}
\section{Request syntax}{
\preformatted{svc$get_job_bookmark(
  JobName = "string",
  RunId = "string"
)
}
}

\keyword{internal}
