% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_tags}
\alias{glue_get_tags}
\title{Retrieves a list of tags associated with a resource}
\usage{
glue_get_tags(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon ARN of the resource for which to retrieve tags.}
}
\description{
Retrieves a list of tags associated with a resource.
}
\section{Request syntax}{
\preformatted{glue$get_tags(
  ResourceArn = "string"
)
}
}

\keyword{internal}
