% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_workflow_runs}
\alias{glue_get_workflow_runs}
\title{Retrieves metadata for all runs of a given workflow}
\usage{
glue_get_workflow_runs(Name, IncludeGraph, NextToken, MaxResults)
}
\arguments{
\item{Name}{[required] Name of the workflow whose metadata of runs should be returned.}

\item{IncludeGraph}{Specifies whether to include the workflow graph in response or not.}

\item{NextToken}{The maximum size of the response.}

\item{MaxResults}{The maximum number of workflow runs to be included in the response.}
}
\description{
Retrieves metadata for all runs of a given workflow.
}
\section{Request syntax}{
\preformatted{svc$get_workflow_runs(
  Name = "string",
  IncludeGraph = TRUE|FALSE,
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
