% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_import_catalog_to_glue}
\alias{glue_import_catalog_to_glue}
\title{Imports an existing Athena Data Catalog to AWS Glue}
\usage{
glue_import_catalog_to_glue(CatalogId)
}
\arguments{
\item{CatalogId}{The ID of the catalog to import. Currently, this should be the AWS
account ID.}
}
\description{
Imports an existing Athena Data Catalog to AWS Glue
}
\section{Request syntax}{
\preformatted{glue$import_catalog_to_glue(
  CatalogId = "string"
)
}
}

\keyword{internal}
