% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_resume_workflow_run}
\alias{glue_resume_workflow_run}
\title{Restarts selected nodes of a previous partially completed workflow run
and resumes the workflow run}
\usage{
glue_resume_workflow_run(Name, RunId, NodeIds)
}
\arguments{
\item{Name}{[required] The name of the workflow to resume.}

\item{RunId}{[required] The ID of the workflow run to resume.}

\item{NodeIds}{[required] A list of the node IDs for the nodes you want to restart. The nodes that
are to be restarted must have a run attempt in the original run.}
}
\description{
Restarts selected nodes of a previous partially completed workflow run
and resumes the workflow run. The selected nodes and all nodes that are
downstream from the selected nodes are run.
}
\section{Request syntax}{
\preformatted{svc$resume_workflow_run(
  Name = "string",
  RunId = "string",
  NodeIds = list(
    "string"
  )
)
}
}

\keyword{internal}
