% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_classifier}
\alias{glue_update_classifier}
\title{Modifies an existing classifier (a GrokClassifier, XMLClassifier, or
JsonClassifier, depending on which field is present)}
\usage{
glue_update_classifier(GrokClassifier, XMLClassifier, JsonClassifier)
}
\arguments{
\item{GrokClassifier}{A \code{GrokClassifier} object with updated fields.}

\item{XMLClassifier}{An \code{XMLClassifier} object with updated fields.}

\item{JsonClassifier}{A \code{JsonClassifier} object with updated fields.}
}
\description{
Modifies an existing classifier (a \code{GrokClassifier}, \code{XMLClassifier}, or
\code{JsonClassifier}, depending on which field is present).
}
\section{Request syntax}{
\preformatted{glue$update_classifier(
  GrokClassifier = list(
    Name = "string",
    Classification = "string",
    GrokPattern = "string",
    CustomPatterns = "string"
  ),
  XMLClassifier = list(
    Name = "string",
    Classification = "string",
    RowTag = "string"
  ),
  JsonClassifier = list(
    Name = "string",
    JsonPath = "string"
  )
)
}
}

\keyword{internal}
