% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_workflow}
\alias{glue_update_workflow}
\title{Updates an existing workflow}
\usage{
glue_update_workflow(Name, Description, DefaultRunProperties)
}
\arguments{
\item{Name}{[required] Name of the workflow to be updated.}

\item{Description}{The description of the workflow.}

\item{DefaultRunProperties}{A collection of properties to be used as part of each execution of the
workflow.}
}
\description{
Updates an existing workflow.
}
\section{Request syntax}{
\preformatted{svc$update_workflow(
  Name = "string",
  Description = "string",
  DefaultRunProperties = list(
    "string"
  )
)
}
}

\keyword{internal}
