% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_batch_delete_recipe_version}
\alias{gluedatabrew_batch_delete_recipe_version}
\title{Deletes one or more versions of a recipe at a time}
\usage{
gluedatabrew_batch_delete_recipe_version(Name, RecipeVersions)
}
\arguments{
\item{Name}{[required] The name of the recipe to be modified.}

\item{RecipeVersions}{[required] An array of version identifiers to be deleted.}
}
\description{
Deletes one or more versions of a recipe at a time.
}
\section{Request syntax}{
\preformatted{svc$batch_delete_recipe_version(
  Name = "string",
  RecipeVersions = list(
    "string"
  )
)
}
}

\keyword{internal}
