% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_create_dataset}
\alias{gluedatabrew_create_dataset}
\title{Creates a new AWS Glue DataBrew dataset for this AWS account}
\usage{
gluedatabrew_create_dataset(Name, FormatOptions, Input, Tags)
}
\arguments{
\item{Name}{[required] The name of the dataset to be created.}

\item{FormatOptions}{}

\item{Input}{[required]}

\item{Tags}{Metadata tags to apply to this dataset.}
}
\description{
Creates a new AWS Glue DataBrew dataset for this AWS account.
}
\section{Request syntax}{
\preformatted{svc$create_dataset(
  Name = "string",
  FormatOptions = list(
    Json = list(
      MultiLine = TRUE|FALSE
    ),
    Excel = list(
      SheetNames = list(
        "string"
      ),
      SheetIndexes = list(
        123
      )
    )
  ),
  Input = list(
    S3InputDefinition = list(
      Bucket = "string",
      Key = "string"
    ),
    DataCatalogInputDefinition = list(
      CatalogId = "string",
      DatabaseName = "string",
      TableName = "string",
      TempDirectory = list(
        Bucket = "string",
        Key = "string"
      )
    )
  ),
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
