% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_create_profile_job}
\alias{gluedatabrew_create_profile_job}
\title{Creates a new job to profile an AWS Glue DataBrew dataset that exists in
the current AWS account}
\usage{
gluedatabrew_create_profile_job(DatasetName, EncryptionKeyArn,
  EncryptionMode, Name, LogSubscription, MaxCapacity, MaxRetries,
  OutputLocation, RoleArn, Tags, Timeout)
}
\arguments{
\item{DatasetName}{[required] The name of the dataset that this job is to act upon.}

\item{EncryptionKeyArn}{The Amazon Resource Name (ARN) of an encryption key that is used to
protect the job.}

\item{EncryptionMode}{The encryption mode for the job, which can be one of the following:
\itemize{
\item \code{SSE-KMS} - para>\code{SSE-KMS} - server-side encryption with AWS
KMS-managed keys.
\item \code{SSE-S3} - Server-side encryption with keys managed by Amazon S3.
}}

\item{Name}{[required] The name of the job to be created.}

\item{LogSubscription}{A value that enables or disables Amazon CloudWatch logging for the
current AWS account. If logging is enabled, CloudWatch writes one log
stream for each job run.}

\item{MaxCapacity}{The maximum number of nodes that DataBrew can use when the job processes
data.}

\item{MaxRetries}{The maximum number of times to retry the job after a job run fails.}

\item{OutputLocation}{[required]}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Access Management
(IAM) role to be assumed for this request.}

\item{Tags}{Metadata tags to apply to this job.}

\item{Timeout}{The job's timeout in minutes. A job that attempts to run longer than
this timeout period ends with a status of \code{TIMEOUT}.}
}
\description{
Creates a new job to profile an AWS Glue DataBrew dataset that exists in
the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$create_profile_job(
  DatasetName = "string",
  EncryptionKeyArn = "string",
  EncryptionMode = "SSE-KMS"|"SSE-S3",
  Name = "string",
  LogSubscription = "ENABLE"|"DISABLE",
  MaxCapacity = 123,
  MaxRetries = 123,
  OutputLocation = list(
    Bucket = "string",
    Key = "string"
  ),
  RoleArn = "string",
  Tags = list(
    "string"
  ),
  Timeout = 123
)
}
}

\keyword{internal}
