% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_create_project}
\alias{gluedatabrew_create_project}
\title{Creates a new AWS Glue DataBrew project in the current AWS account}
\usage{
gluedatabrew_create_project(DatasetName, Name, RecipeName, Sample,
  RoleArn, Tags)
}
\arguments{
\item{DatasetName}{[required] The name of the dataset to associate this project with.}

\item{Name}{[required] A unique name for the new project.}

\item{RecipeName}{[required] The name of an existing recipe to associate with the project.}

\item{Sample}{}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Access Management
(IAM) role to be assumed for this request.}

\item{Tags}{Metadata tags to apply to this project.}
}
\description{
Creates a new AWS Glue DataBrew project in the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$create_project(
  DatasetName = "string",
  Name = "string",
  RecipeName = "string",
  Sample = list(
    Size = 123,
    Type = "FIRST_N"|"LAST_N"|"RANDOM"
  ),
  RoleArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
