% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_create_recipe_job}
\alias{gluedatabrew_create_recipe_job}
\title{Creates a new job for an existing AWS Glue DataBrew recipe in the
current AWS account}
\usage{
gluedatabrew_create_recipe_job(DatasetName, EncryptionKeyArn,
  EncryptionMode, Name, LogSubscription, MaxCapacity, MaxRetries, Outputs,
  ProjectName, RecipeReference, RoleArn, Tags, Timeout)
}
\arguments{
\item{DatasetName}{The name of the dataset that this job processes.}

\item{EncryptionKeyArn}{The Amazon Resource Name (ARN) of an encryption key that is used to
protect the job.}

\item{EncryptionMode}{The encryption mode for the job, which can be one of the following:
\itemize{
\item \code{SSE-KMS} - Server-side encryption with AWS KMS-managed keys.
\item \code{SSE-S3} - Server-side encryption with keys managed by Amazon S3.
}}

\item{Name}{[required] A unique name for the job.}

\item{LogSubscription}{A value that enables or disables Amazon CloudWatch logging for the
current AWS account. If logging is enabled, CloudWatch writes one log
stream for each job run.}

\item{MaxCapacity}{The maximum number of nodes that DataBrew can consume when the job
processes data.}

\item{MaxRetries}{The maximum number of times to retry the job after a job run fails.}

\item{Outputs}{[required] One or more artifacts that represent the output from running the job.}

\item{ProjectName}{Either the name of an existing project, or a combination of a recipe and
a dataset to associate with the recipe.}

\item{RecipeReference}{}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Access Management
(IAM) role to be assumed for this request.}

\item{Tags}{Metadata tags to apply to this job dataset.}

\item{Timeout}{The job's timeout in minutes. A job that attempts to run longer than
this timeout period ends with a status of \code{TIMEOUT}.}
}
\description{
Creates a new job for an existing AWS Glue DataBrew recipe in the
current AWS account. You can create a standalone job using either a
project, or a combination of a recipe and a dataset.
}
\section{Request syntax}{
\preformatted{svc$create_recipe_job(
  DatasetName = "string",
  EncryptionKeyArn = "string",
  EncryptionMode = "SSE-KMS"|"SSE-S3",
  Name = "string",
  LogSubscription = "ENABLE"|"DISABLE",
  MaxCapacity = 123,
  MaxRetries = 123,
  Outputs = list(
    list(
      CompressionFormat = "GZIP"|"LZ4"|"SNAPPY"|"BZIP2"|"DEFLATE"|"LZO"|"BROTLI"|"ZSTD"|"ZLIB",
      Format = "CSV"|"JSON"|"PARQUET"|"GLUEPARQUET"|"AVRO"|"ORC"|"XML",
      PartitionColumns = list(
        "string"
      ),
      Location = list(
        Bucket = "string",
        Key = "string"
      ),
      Overwrite = TRUE|FALSE
    )
  ),
  ProjectName = "string",
  RecipeReference = list(
    Name = "string",
    RecipeVersion = "string"
  ),
  RoleArn = "string",
  Tags = list(
    "string"
  ),
  Timeout = 123
)
}
}

\keyword{internal}
