% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_create_schedule}
\alias{gluedatabrew_create_schedule}
\title{Creates a new schedule for one or more AWS Glue DataBrew jobs}
\usage{
gluedatabrew_create_schedule(JobNames, CronExpression, Tags, Name)
}
\arguments{
\item{JobNames}{The name or names of one or more jobs to be run.}

\item{CronExpression}{[required] The date or dates and time or times, in \code{cron} format, when the jobs are
to be run.}

\item{Tags}{Metadata tags to apply to this schedule.}

\item{Name}{[required] A unique name for the schedule.}
}
\description{
Creates a new schedule for one or more AWS Glue DataBrew jobs. Jobs can
be run at a specific date and time, or at regular intervals.
}
\section{Request syntax}{
\preformatted{svc$create_schedule(
  JobNames = list(
    "string"
  ),
  CronExpression = "string",
  Tags = list(
    "string"
  ),
  Name = "string"
)
}
}

\keyword{internal}
