% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_delete_project}
\alias{gluedatabrew_delete_project}
\title{Deletes an existing AWS Glue DataBrew project from the current AWS
account}
\usage{
gluedatabrew_delete_project(Name)
}
\arguments{
\item{Name}{[required] The name of the project to be deleted.}
}
\description{
Deletes an existing AWS Glue DataBrew project from the current AWS
account.
}
\section{Request syntax}{
\preformatted{svc$delete_project(
  Name = "string"
)
}
}

\keyword{internal}
