% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_delete_schedule}
\alias{gluedatabrew_delete_schedule}
\title{Deletes the specified AWS Glue DataBrew schedule from the current AWS
account}
\usage{
gluedatabrew_delete_schedule(Name)
}
\arguments{
\item{Name}{[required] The name of the schedule to be deleted.}
}
\description{
Deletes the specified AWS Glue DataBrew schedule from the current AWS
account.
}
\section{Request syntax}{
\preformatted{svc$delete_schedule(
  Name = "string"
)
}
}

\keyword{internal}
