% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_describe_job}
\alias{gluedatabrew_describe_job}
\title{Returns the definition of a specific AWS Glue DataBrew job that is in
the current AWS account}
\usage{
gluedatabrew_describe_job(Name)
}
\arguments{
\item{Name}{[required] The name of the job to be described.}
}
\description{
Returns the definition of a specific AWS Glue DataBrew job that is in
the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$describe_job(
  Name = "string"
)
}
}

\keyword{internal}
