% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_describe_project}
\alias{gluedatabrew_describe_project}
\title{Returns the definition of a specific AWS Glue DataBrew project that is
in the current AWS account}
\usage{
gluedatabrew_describe_project(Name)
}
\arguments{
\item{Name}{[required] The name of the project to be described.}
}
\description{
Returns the definition of a specific AWS Glue DataBrew project that is
in the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$describe_project(
  Name = "string"
)
}
}

\keyword{internal}
