% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_describe_recipe}
\alias{gluedatabrew_describe_recipe}
\title{Returns the definition of a specific AWS Glue DataBrew recipe that is in
the current AWS account}
\usage{
gluedatabrew_describe_recipe(Name, RecipeVersion)
}
\arguments{
\item{Name}{[required] The name of the recipe to be described.}

\item{RecipeVersion}{The recipe version identifier. If this parameter isn't specified, then
the latest published version is returned.}
}
\description{
Returns the definition of a specific AWS Glue DataBrew recipe that is in
the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$describe_recipe(
  Name = "string",
  RecipeVersion = "string"
)
}
}

\keyword{internal}
