% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_job_runs}
\alias{gluedatabrew_list_job_runs}
\title{Lists all of the previous runs of a particular AWS Glue DataBrew job in
the current AWS account}
\usage{
gluedatabrew_list_job_runs(Name, MaxResults, NextToken)
}
\arguments{
\item{Name}{[required] The name of the job.}

\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{A token generated by AWS Glue DataBrew that specifies where to continue
pagination if a previous request was truncated. To get the next set of
pages, pass in the NextToken value from the response object of the
previous page call.}
}
\description{
Lists all of the previous runs of a particular AWS Glue DataBrew job in
the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_job_runs(
  Name = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
