% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_jobs}
\alias{gluedatabrew_list_jobs}
\title{Lists the AWS Glue DataBrew jobs in the current AWS account}
\usage{
gluedatabrew_list_jobs(DatasetName, MaxResults, NextToken, ProjectName)
}
\arguments{
\item{DatasetName}{The name of a dataset. Using this parameter indicates to return only
those jobs that act on the specified dataset.}

\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{A token generated by DataBrew that specifies where to continue
pagination if a previous request was truncated. To get the next set of
pages, pass in the NextToken value from the response object of the
previous page call.}

\item{ProjectName}{The name of a project. Using this parameter indicates to return only
those jobs that are associated with the specified project.}
}
\description{
Lists the AWS Glue DataBrew jobs in the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_jobs(
  DatasetName = "string",
  MaxResults = 123,
  NextToken = "string",
  ProjectName = "string"
)
}
}

\keyword{internal}
