% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_projects}
\alias{gluedatabrew_list_projects}
\title{Lists all of the DataBrew projects in the current AWS account}
\usage{
gluedatabrew_list_projects(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return in this request.}
}
\description{
Lists all of the DataBrew projects in the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_projects(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
