% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_recipe_versions}
\alias{gluedatabrew_list_recipe_versions}
\title{Lists all of the versions of a particular AWS Glue DataBrew recipe in
the current AWS account}
\usage{
gluedatabrew_list_recipe_versions(MaxResults, NextToken, Name)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{Name}{[required] The name of the recipe for which to return version information.}
}
\description{
Lists all of the versions of a particular AWS Glue DataBrew recipe in
the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_recipe_versions(
  MaxResults = 123,
  NextToken = "string",
  Name = "string"
)
}
}

\keyword{internal}
