% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_schedules}
\alias{gluedatabrew_list_schedules}
\title{Lists the AWS Glue DataBrew schedules in the current AWS account}
\usage{
gluedatabrew_list_schedules(JobName, MaxResults, NextToken)
}
\arguments{
\item{JobName}{The name of the job that these schedules apply to.}

\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}
}
\description{
Lists the AWS Glue DataBrew schedules in the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_schedules(
  JobName = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
