% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_tags_for_resource}
\alias{gluedatabrew_list_tags_for_resource}
\title{Lists all the tags for an AWS Glue DataBrew resource}
\usage{
gluedatabrew_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the
DataBrew resource.}
}
\description{
Lists all the tags for an AWS Glue DataBrew resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
