% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_publish_recipe}
\alias{gluedatabrew_publish_recipe}
\title{Publishes a new major version of an AWS Glue DataBrew recipe that exists
in the current AWS account}
\usage{
gluedatabrew_publish_recipe(Description, Name)
}
\arguments{
\item{Description}{A description of the recipe to be published, for this version of the
recipe.}

\item{Name}{[required] The name of the recipe to be published.}
}
\description{
Publishes a new major version of an AWS Glue DataBrew recipe that exists
in the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$publish_recipe(
  Description = "string",
  Name = "string"
)
}
}

\keyword{internal}
