% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_start_job_run}
\alias{gluedatabrew_start_job_run}
\title{Runs an AWS Glue DataBrew job that exists in the current AWS account}
\usage{
gluedatabrew_start_job_run(Name)
}
\arguments{
\item{Name}{[required] The name of the job to be run.}
}
\description{
Runs an AWS Glue DataBrew job that exists in the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$start_job_run(
  Name = "string"
)
}
}

\keyword{internal}
