% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_start_project_session}
\alias{gluedatabrew_start_project_session}
\title{Creates an interactive session, enabling you to manipulate an AWS Glue
DataBrew project}
\usage{
gluedatabrew_start_project_session(Name, AssumeControl)
}
\arguments{
\item{Name}{[required] The name of the project to act upon.}

\item{AssumeControl}{A value that, if true, enables you to take control of a session, even if
a different client is currently accessing the project.}
}
\description{
Creates an interactive session, enabling you to manipulate an AWS Glue
DataBrew project.
}
\section{Request syntax}{
\preformatted{svc$start_project_session(
  Name = "string",
  AssumeControl = TRUE|FALSE
)
}
}

\keyword{internal}
