% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_stop_job_run}
\alias{gluedatabrew_stop_job_run}
\title{Stops the specified job from running in the current AWS account}
\usage{
gluedatabrew_stop_job_run(Name, RunId)
}
\arguments{
\item{Name}{[required] The name of the job to be stopped.}

\item{RunId}{[required] The ID of the job run to be stopped.}
}
\description{
Stops the specified job from running in the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$stop_job_run(
  Name = "string",
  RunId = "string"
)
}
}

\keyword{internal}
