% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_untag_resource}
\alias{gluedatabrew_untag_resource}
\title{Removes metadata tags from an AWS Glue DataBrew resource}
\usage{
gluedatabrew_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] An DataBrew resource from which you want to remove a tag or tags. The
value for this parameter is an Amazon Resource Name (ARN).}

\item{TagKeys}{[required] The tag keys (names) of one or more tags to be removed.}
}
\description{
Removes metadata tags from an AWS Glue DataBrew resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
