% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_update_project}
\alias{gluedatabrew_update_project}
\title{Modifies the definition of an existing AWS Glue DataBrew project in the
current AWS account}
\usage{
gluedatabrew_update_project(Sample, RoleArn, Name)
}
\arguments{
\item{Sample}{}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role to be assumed for this
request.}

\item{Name}{[required] The name of the project to be updated.}
}
\description{
Modifies the definition of an existing AWS Glue DataBrew project in the
current AWS account.
}
\section{Request syntax}{
\preformatted{svc$update_project(
  Sample = list(
    Size = 123,
    Type = "FIRST_N"|"LAST_N"|"RANDOM"
  ),
  RoleArn = "string",
  Name = "string"
)
}
}

\keyword{internal}
