% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_update_recipe}
\alias{gluedatabrew_update_recipe}
\title{Modifies the definition of the latest working version of an AWS Glue
DataBrew recipe in the current AWS account}
\usage{
gluedatabrew_update_recipe(Description, Name, Steps)
}
\arguments{
\item{Description}{A description of the recipe.}

\item{Name}{[required] The name of the recipe to be updated.}

\item{Steps}{One or more steps to be performed by the recipe. Each step consists of
an action, and the conditions under which the action should succeed.}
}
\description{
Modifies the definition of the latest working version of an AWS Glue
DataBrew recipe in the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$update_recipe(
  Description = "string",
  Name = "string",
  Steps = list(
    list(
      Action = list(
        Operation = "string",
        Parameters = list(
          "string"
        )
      ),
      ConditionExpressions = list(
        list(
          Condition = "string",
          Value = "string",
          TargetColumn = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
