% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_update_recipe_job}
\alias{gluedatabrew_update_recipe_job}
\title{Modifies the definition of an existing AWS Glue DataBrew recipe job in
the current AWS account}
\usage{
gluedatabrew_update_recipe_job(EncryptionKeyArn, EncryptionMode, Name,
  LogSubscription, MaxCapacity, MaxRetries, Outputs, RoleArn, Timeout)
}
\arguments{
\item{EncryptionKeyArn}{The Amazon Resource Name (ARN) of an encryption key that is used to
protect the job.}

\item{EncryptionMode}{The encryption mode for the job, which can be one of the following:
\itemize{
\item \code{SSE-KMS} - Server-side encryption with AWS KMS-managed keys.
\item \code{SSE-S3} - Server-side encryption with keys managed by Amazon S3.
}}

\item{Name}{[required] The name of the job to update.}

\item{LogSubscription}{A value that enables or disables Amazon CloudWatch logging for the
current AWS account. If logging is enabled, CloudWatch writes one log
stream for each job run.}

\item{MaxCapacity}{The maximum number of nodes that DataBrew can consume when the job
processes data.}

\item{MaxRetries}{The maximum number of times to retry the job after a job run fails.}

\item{Outputs}{[required] One or more artifacts that represent the output from running the job.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Access Management
(IAM) role to be assumed for this request.}

\item{Timeout}{The job's timeout in minutes. A job that attempts to run longer than
this timeout period ends with a status of \code{TIMEOUT}.}
}
\description{
Modifies the definition of an existing AWS Glue DataBrew recipe job in
the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$update_recipe_job(
  EncryptionKeyArn = "string",
  EncryptionMode = "SSE-KMS"|"SSE-S3",
  Name = "string",
  LogSubscription = "ENABLE"|"DISABLE",
  MaxCapacity = 123,
  MaxRetries = 123,
  Outputs = list(
    list(
      CompressionFormat = "GZIP"|"LZ4"|"SNAPPY"|"BZIP2"|"DEFLATE"|"LZO"|"BROTLI"|"ZSTD"|"ZLIB",
      Format = "CSV"|"JSON"|"PARQUET"|"GLUEPARQUET"|"AVRO"|"ORC"|"XML",
      PartitionColumns = list(
        "string"
      ),
      Location = list(
        Bucket = "string",
        Key = "string"
      ),
      Overwrite = TRUE|FALSE
    )
  ),
  RoleArn = "string",
  Timeout = 123
)
}
}

\keyword{internal}
