% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_update_schedule}
\alias{gluedatabrew_update_schedule}
\title{Modifies the definition of an existing AWS Glue DataBrew schedule in the
current AWS account}
\usage{
gluedatabrew_update_schedule(JobNames, CronExpression, Name)
}
\arguments{
\item{JobNames}{The name or names of one or more jobs to be run for this schedule.}

\item{CronExpression}{[required] The date or dates and time or times, in \code{cron} format, when the jobs are
to be run.}

\item{Name}{[required] The name of the schedule to update.}
}
\description{
Modifies the definition of an existing AWS Glue DataBrew schedule in the
current AWS account.
}
\section{Request syntax}{
\preformatted{svc$update_schedule(
  JobNames = list(
    "string"
  ),
  CronExpression = "string",
  Name = "string"
)
}
}

\keyword{internal}
