% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_connector_definition}
\alias{greengrass_create_connector_definition}
\title{Creates a connector definition}
\usage{
greengrass_create_connector_definition(AmznClientToken, InitialVersion,
  Name, tags)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{InitialVersion}{Information about the initial version of the connector definition.}

\item{Name}{The name of the connector definition.}

\item{tags}{Tag(s) to add to the new resource.}
}
\description{
Creates a connector definition. You may provide the initial version of the connector definition now or use ''CreateConnectorDefinitionVersion'' at a later time.
}
\section{Request syntax}{
\preformatted{svc$create_connector_definition(
  AmznClientToken = "string",
  InitialVersion = list(
    Connectors = list(
      list(
        ConnectorArn = "string",
        Id = "string",
        Parameters = list(
          "string"
        )
      )
    )
  ),
  Name = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
