% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_connector_definition_version}
\alias{greengrass_create_connector_definition_version}
\title{Creates a version of a connector definition which has already been
defined}
\usage{
greengrass_create_connector_definition_version(AmznClientToken,
  ConnectorDefinitionId, Connectors)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}

\item{Connectors}{A list of references to connectors in this version, with their corresponding configuration settings.}
}
\description{
Creates a version of a connector definition which has already been defined.
}
\section{Request syntax}{
\preformatted{greengrass$create_connector_definition_version(
  AmznClientToken = "string",
  ConnectorDefinitionId = "string",
  Connectors = list(
    list(
      ConnectorArn = "string",
      Id = "string",
      Parameters = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
