% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_function_definition}
\alias{greengrass_create_function_definition}
\title{Creates a Lambda function definition which contains a list of Lambda
functions and their configurations to be used in a group}
\usage{
greengrass_create_function_definition(AmznClientToken, InitialVersion,
  Name, tags)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{InitialVersion}{Information about the initial version of the function definition.}

\item{Name}{The name of the function definition.}

\item{tags}{Tag(s) to add to the new resource.}
}
\description{
Creates a Lambda function definition which contains a list of Lambda functions and their configurations to be used in a group. You can create an initial version of the definition by providing a list of Lambda functions and their configurations now, or use ''CreateFunctionDefinitionVersion'' later.
}
\section{Request syntax}{
\preformatted{svc$create_function_definition(
  AmznClientToken = "string",
  InitialVersion = list(
    DefaultConfig = list(
      Execution = list(
        IsolationMode = "GreengrassContainer"|"NoContainer",
        RunAs = list(
          Gid = 123,
          Uid = 123
        )
      )
    ),
    Functions = list(
      list(
        FunctionArn = "string",
        FunctionConfiguration = list(
          EncodingType = "binary"|"json",
          Environment = list(
            AccessSysfs = TRUE|FALSE,
            Execution = list(
              IsolationMode = "GreengrassContainer"|"NoContainer",
              RunAs = list(
                Gid = 123,
                Uid = 123
              )
            ),
            ResourceAccessPolicies = list(
              list(
                Permission = "ro"|"rw",
                ResourceId = "string"
              )
            ),
            Variables = list(
              "string"
            )
          ),
          ExecArgs = "string",
          Executable = "string",
          MemorySize = 123,
          Pinned = TRUE|FALSE,
          Timeout = 123
        ),
        Id = "string"
      )
    )
  ),
  Name = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
