% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_group_version}
\alias{greengrass_create_group_version}
\title{Creates a version of a group which has already been defined}
\usage{
greengrass_create_group_version(AmznClientToken,
  ConnectorDefinitionVersionArn, CoreDefinitionVersionArn,
  DeviceDefinitionVersionArn, FunctionDefinitionVersionArn, GroupId,
  LoggerDefinitionVersionArn, ResourceDefinitionVersionArn,
  SubscriptionDefinitionVersionArn)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{ConnectorDefinitionVersionArn}{The ARN of the connector definition version for this group.}

\item{CoreDefinitionVersionArn}{The ARN of the core definition version for this group.}

\item{DeviceDefinitionVersionArn}{The ARN of the device definition version for this group.}

\item{FunctionDefinitionVersionArn}{The ARN of the function definition version for this group.}

\item{GroupId}{[required] The ID of the Greengrass group.}

\item{LoggerDefinitionVersionArn}{The ARN of the logger definition version for this group.}

\item{ResourceDefinitionVersionArn}{The ARN of the resource definition version for this group.}

\item{SubscriptionDefinitionVersionArn}{The ARN of the subscription definition version for this group.}
}
\description{
Creates a version of a group which has already been defined.
}
\section{Request syntax}{
\preformatted{greengrass$create_group_version(
  AmznClientToken = "string",
  ConnectorDefinitionVersionArn = "string",
  CoreDefinitionVersionArn = "string",
  DeviceDefinitionVersionArn = "string",
  FunctionDefinitionVersionArn = "string",
  GroupId = "string",
  LoggerDefinitionVersionArn = "string",
  ResourceDefinitionVersionArn = "string",
  SubscriptionDefinitionVersionArn = "string"
)
}
}

\keyword{internal}
