% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_resource_definition}
\alias{greengrass_create_resource_definition}
\title{Creates a resource definition which contains a list of resources to be
used in a group}
\usage{
greengrass_create_resource_definition(AmznClientToken, InitialVersion,
  Name, tags)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{InitialVersion}{Information about the initial version of the resource definition.}

\item{Name}{The name of the resource definition.}

\item{tags}{Tag(s) to add to the new resource}
}
\description{
Creates a resource definition which contains a list of resources to be used in a group. You can create an initial version of the definition by providing a list of resources now, or use ''CreateResourceDefinitionVersion'' later.
}
\section{Request syntax}{
\preformatted{svc$create_resource_definition(
  AmznClientToken = "string",
  InitialVersion = list(
    Resources = list(
      list(
        Id = "string",
        Name = "string",
        ResourceDataContainer = list(
          LocalDeviceResourceData = list(
            GroupOwnerSetting = list(
              AutoAddGroupOwner = TRUE|FALSE,
              GroupOwner = "string"
            ),
            SourcePath = "string"
          ),
          LocalVolumeResourceData = list(
            DestinationPath = "string",
            GroupOwnerSetting = list(
              AutoAddGroupOwner = TRUE|FALSE,
              GroupOwner = "string"
            ),
            SourcePath = "string"
          ),
          S3MachineLearningModelResourceData = list(
            DestinationPath = "string",
            S3Uri = "string"
          ),
          SageMakerMachineLearningModelResourceData = list(
            DestinationPath = "string",
            SageMakerJobArn = "string"
          ),
          SecretsManagerSecretResourceData = list(
            ARN = "string",
            AdditionalStagingLabelsToDownload = list(
              "string"
            )
          )
        )
      )
    )
  ),
  Name = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
