% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_resource_definition_version}
\alias{greengrass_create_resource_definition_version}
\title{Creates a version of a resource definition that has already been defined}
\usage{
greengrass_create_resource_definition_version(AmznClientToken,
  ResourceDefinitionId, Resources)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{ResourceDefinitionId}{[required] The ID of the resource definition.}

\item{Resources}{A list of resources.}
}
\description{
Creates a version of a resource definition that has already been defined.
}
\section{Request syntax}{
\preformatted{svc$create_resource_definition_version(
  AmznClientToken = "string",
  ResourceDefinitionId = "string",
  Resources = list(
    list(
      Id = "string",
      Name = "string",
      ResourceDataContainer = list(
        LocalDeviceResourceData = list(
          GroupOwnerSetting = list(
            AutoAddGroupOwner = TRUE|FALSE,
            GroupOwner = "string"
          ),
          SourcePath = "string"
        ),
        LocalVolumeResourceData = list(
          DestinationPath = "string",
          GroupOwnerSetting = list(
            AutoAddGroupOwner = TRUE|FALSE,
            GroupOwner = "string"
          ),
          SourcePath = "string"
        ),
        S3MachineLearningModelResourceData = list(
          DestinationPath = "string",
          S3Uri = "string"
        ),
        SageMakerMachineLearningModelResourceData = list(
          DestinationPath = "string",
          SageMakerJobArn = "string"
        ),
        SecretsManagerSecretResourceData = list(
          ARN = "string",
          AdditionalStagingLabelsToDownload = list(
            "string"
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
