% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_delete_core_definition}
\alias{greengrass_delete_core_definition}
\title{Deletes a core definition}
\usage{
greengrass_delete_core_definition(CoreDefinitionId)
}
\arguments{
\item{CoreDefinitionId}{[required] The ID of the core definition.}
}
\description{
Deletes a core definition.
}
\section{Request syntax}{
\preformatted{greengrass$delete_core_definition(
  CoreDefinitionId = "string"
)
}
}

\keyword{internal}
