% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_disassociate_service_role_from_account}
\alias{greengrass_disassociate_service_role_from_account}
\title{Disassociates the service role from your account}
\usage{
greengrass_disassociate_service_role_from_account()
}
\description{
Disassociates the service role from your account. Without a service role, deployments will not work.
}
\section{Request syntax}{
\preformatted{greengrass$disassociate_service_role_from_account()
}
}

\keyword{internal}
