% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_connector_definition}
\alias{greengrass_get_connector_definition}
\title{Retrieves information about a connector definition}
\usage{
greengrass_get_connector_definition(ConnectorDefinitionId)
}
\arguments{
\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}
}
\description{
Retrieves information about a connector definition.
}
\section{Request syntax}{
\preformatted{greengrass$get_connector_definition(
  ConnectorDefinitionId = "string"
)
}
}

\keyword{internal}
