% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_core_definition_version}
\alias{greengrass_get_core_definition_version}
\title{Retrieves information about a core definition version}
\usage{
greengrass_get_core_definition_version(CoreDefinitionId,
  CoreDefinitionVersionId)
}
\arguments{
\item{CoreDefinitionId}{[required] The ID of the core definition.}

\item{CoreDefinitionVersionId}{[required] The ID of the core definition version.}
}
\description{
Retrieves information about a core definition version.
}
\section{Request syntax}{
\preformatted{greengrass$get_core_definition_version(
  CoreDefinitionId = "string",
  CoreDefinitionVersionId = "string"
)
}
}

\keyword{internal}
