% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_logger_definition_version}
\alias{greengrass_get_logger_definition_version}
\title{Retrieves information about a logger definition version}
\usage{
greengrass_get_logger_definition_version(LoggerDefinitionId,
  LoggerDefinitionVersionId, NextToken)
}
\arguments{
\item{LoggerDefinitionId}{[required] The ID of the logger definition.}

\item{LoggerDefinitionVersionId}{[required] The ID of the logger definition version.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Retrieves information about a logger definition version.
}
\section{Request syntax}{
\preformatted{greengrass$get_logger_definition_version(
  LoggerDefinitionId = "string",
  LoggerDefinitionVersionId = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
