% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_resource_definition_version}
\alias{greengrass_get_resource_definition_version}
\title{Retrieves information about a resource definition version, including
which resources are included in the version}
\usage{
greengrass_get_resource_definition_version(ResourceDefinitionId,
  ResourceDefinitionVersionId)
}
\arguments{
\item{ResourceDefinitionId}{[required] The ID of the resource definition.}

\item{ResourceDefinitionVersionId}{[required] The ID of the resource definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListResourceDefinitionVersions'' requests. If the version is the last one that was associated with a resource definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.}
}
\description{
Retrieves information about a resource definition version, including which resources are included in the version.
}
\section{Request syntax}{
\preformatted{svc$get_resource_definition_version(
  ResourceDefinitionId = "string",
  ResourceDefinitionVersionId = "string"
)
}
}

\keyword{internal}
