% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_thing_runtime_configuration}
\alias{greengrass_get_thing_runtime_configuration}
\title{Get the runtime configuration of a thing}
\usage{
greengrass_get_thing_runtime_configuration(ThingName)
}
\arguments{
\item{ThingName}{[required] The thing name.}
}
\description{
Get the runtime configuration of a thing.
}
\section{Request syntax}{
\preformatted{svc$get_thing_runtime_configuration(
  ThingName = "string"
)
}
}

\keyword{internal}
