% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_bulk_deployment_detailed_reports}
\alias{greengrass_list_bulk_deployment_detailed_reports}
\title{Gets a paginated list of the deployments that have been started in a
bulk deployment operation, and their current deployment status}
\usage{
greengrass_list_bulk_deployment_detailed_reports(BulkDeploymentId,
  MaxResults, NextToken)
}
\arguments{
\item{BulkDeploymentId}{[required] The ID of the bulk deployment.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Gets a paginated list of the deployments that have been started in a bulk deployment operation, and their current deployment status.
}
\section{Request syntax}{
\preformatted{greengrass$list_bulk_deployment_detailed_reports(
  BulkDeploymentId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
