% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_reset_deployments}
\alias{greengrass_reset_deployments}
\title{Resets a group's deployments}
\usage{
greengrass_reset_deployments(AmznClientToken, Force, GroupId)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{Force}{If true, performs a best-effort only core reset.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Resets a group's deployments.
}
\section{Request syntax}{
\preformatted{svc$reset_deployments(
  AmznClientToken = "string",
  Force = TRUE|FALSE,
  GroupId = "string"
)
}
}

\keyword{internal}
