% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_connector_definition}
\alias{greengrass_update_connector_definition}
\title{Updates a connector definition}
\usage{
greengrass_update_connector_definition(ConnectorDefinitionId, Name)
}
\arguments{
\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a connector definition.
}
\section{Request syntax}{
\preformatted{svc$update_connector_definition(
  ConnectorDefinitionId = "string",
  Name = "string"
)
}
}

\keyword{internal}
