% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_core_definition}
\alias{greengrass_update_core_definition}
\title{Updates a core definition}
\usage{
greengrass_update_core_definition(CoreDefinitionId, Name)
}
\arguments{
\item{CoreDefinitionId}{[required] The ID of the core definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a core definition.
}
\section{Request syntax}{
\preformatted{greengrass$update_core_definition(
  CoreDefinitionId = "string",
  Name = "string"
)
}
}

\keyword{internal}
