% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_service.R
\name{greengrassv2}
\alias{greengrassv2}
\title{AWS IoT Greengrass V2}
\usage{
greengrassv2(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
AWS IoT Greengrass brings local compute, messaging, data management,
sync, and ML inference capabilities to edge devices. This enables
devices to collect and analyze data closer to the source of information,
react autonomously to local events, and communicate securely with each
other on local networks. Local devices can also communicate securely
with AWS IoT Core and export IoT data to the AWS Cloud. AWS IoT
Greengrass developers can use AWS Lambda functions and components to
create and deploy applications to fleets of edge devices for local
operation.

AWS IoT Greengrass Version 2 provides a new major version of the AWS IoT
Greengrass Core software, new APIs, and a new console. Use this API
reference to learn how to use the AWS IoT Greengrass V2 API operations
to manage components, manage deployments, and core devices.

For more information, see \href{https://docs.aws.amazon.com/greengrass/v2/developerguide/what-is-iot-greengrass.html}{What is AWS IoT Greengrass?}
in the \emph{AWS IoT Greengrass V2 Developer Guide}.
}
\section{Service syntax}{
\preformatted{svc <- greengrassv2(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=greengrassv2_cancel_deployment]{cancel_deployment} \tab Cancels a deployment\cr
\link[=greengrassv2_create_component_version]{create_component_version} \tab Creates a component\cr
\link[=greengrassv2_create_deployment]{create_deployment} \tab Creates a continuous deployment for a target, which is a AWS IoT Greengrass core device or group of core devices\cr
\link[=greengrassv2_delete_component]{delete_component} \tab Deletes a version of a component from AWS IoT Greengrass\cr
\link[=greengrassv2_delete_core_device]{delete_core_device} \tab Deletes a AWS IoT Greengrass core device, which is an AWS IoT thing\cr
\link[=greengrassv2_describe_component]{describe_component} \tab Retrieves metadata for a version of a component\cr
\link[=greengrassv2_get_component]{get_component} \tab Gets the recipe for a version of a component\cr
\link[=greengrassv2_get_component_version_artifact]{get_component_version_artifact} \tab Gets the pre-signed URL to a component artifact in an S3 bucket\cr
\link[=greengrassv2_get_core_device]{get_core_device} \tab Retrieves metadata for a AWS IoT Greengrass core device\cr
\link[=greengrassv2_get_deployment]{get_deployment} \tab Gets a deployment\cr
\link[=greengrassv2_list_components]{list_components} \tab Retrieves a paginated list of component summaries\cr
\link[=greengrassv2_list_component_versions]{list_component_versions} \tab Retrieves a paginated list of all versions for a component\cr
\link[=greengrassv2_list_core_devices]{list_core_devices} \tab Retrieves a paginated list of AWS IoT Greengrass core devices\cr
\link[=greengrassv2_list_deployments]{list_deployments} \tab Retrieves a paginated list of deployments\cr
\link[=greengrassv2_list_effective_deployments]{list_effective_deployments} \tab Retrieves a paginated list of deployment jobs that AWS IoT Greengrass sends to AWS IoT Greengrass core devices\cr
\link[=greengrassv2_list_installed_components]{list_installed_components} \tab Retrieves a paginated list of the components that a AWS IoT Greengrass core device runs\cr
\link[=greengrassv2_list_tags_for_resource]{list_tags_for_resource} \tab Retrieves the list of tags for an AWS IoT Greengrass resource\cr
\link[=greengrassv2_resolve_component_candidates]{resolve_component_candidates} \tab Retrieves a list of components that meet the component, version, and platform requirements of a deployment\cr
\link[=greengrassv2_tag_resource]{tag_resource} \tab Adds tags to an AWS IoT Greengrass resource\cr
\link[=greengrassv2_untag_resource]{untag_resource} \tab Removes a tag from an AWS IoT Greengrass resource
}
}

\examples{
\dontrun{
svc <- greengrassv2()
svc$cancel_deployment(
  Foo = 123
)
}

}
