% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_cancel_deployment}
\alias{greengrassv2_cancel_deployment}
\title{Cancels a deployment}
\usage{
greengrassv2_cancel_deployment(deploymentId)
}
\arguments{
\item{deploymentId}{[required] The ID of the deployment.}
}
\description{
Cancels a deployment. This operation cancels the deployment for devices
that haven't yet received it. If a device already received the
deployment, this operation doesn't change anything for that device.
}
\section{Request syntax}{
\preformatted{svc$cancel_deployment(
  deploymentId = "string"
)
}
}

\keyword{internal}
