% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_create_deployment}
\alias{greengrassv2_create_deployment}
\title{Creates a continuous deployment for a target, which is a AWS IoT
Greengrass core device or group of core devices}
\usage{
greengrassv2_create_deployment(targetArn, deploymentName, components,
  iotJobConfiguration, deploymentPolicies, tags)
}
\arguments{
\item{targetArn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the target AWS IoT thing or thing group.}

\item{deploymentName}{The name of the deployment.

You can create deployments without names. If you create a deployment
without a name, the AWS IoT Greengrass V2 console shows the deployment
name as \verb{<targetType>:<targetName>}, where \code{targetType} and \code{targetName}
are the type and name of the deployment target.}

\item{components}{The components to deploy. This is a dictionary, where each key is the
name of a component, and each key's value is the version and
configuration to deploy for that component.}

\item{iotJobConfiguration}{The job configuration for the deployment configuration. The job
configuration specifies the rollout, timeout, and stop configurations
for the deployment configuration.}

\item{deploymentPolicies}{The deployment policies for the deployment. These policies define how
the deployment updates components and handles failure.}

\item{tags}{A list of key-value pairs that contain metadata for the resource. For
more information, see Tag your resources in the \emph{AWS IoT Greengrass V2
Developer Guide}.}
}
\description{
Creates a continuous deployment for a target, which is a AWS IoT
Greengrass core device or group of core devices. When you add a new core
device to a group of core devices that has a deployment, AWS IoT
Greengrass deploys that group's deployment to the new device.

You can define one deployment for each target. When you create a new
deployment for a target that has an existing deployment, you replace the
previous deployment. AWS IoT Greengrass applies the new deployment to
the target devices.

Every deployment has a revision number that indicates how many
deployment revisions you define for a target. Use this operation to
create a new revision of an existing deployment. This operation returns
the revision number of the new deployment when you create it.

For more information, see the Create deployments in the \emph{AWS IoT
Greengrass V2 Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_deployment(
  targetArn = "string",
  deploymentName = "string",
  components = list(
    list(
      componentVersion = "string",
      configurationUpdate = list(
        merge = "string",
        reset = list(
          "string"
        )
      ),
      runWith = list(
        posixUser = "string"
      )
    )
  ),
  iotJobConfiguration = list(
    jobExecutionsRolloutConfig = list(
      exponentialRate = list(
        baseRatePerMinute = 123,
        incrementFactor = 123.0,
        rateIncreaseCriteria = list(
          numberOfNotifiedThings = 123,
          numberOfSucceededThings = 123
        )
      ),
      maximumPerMinute = 123
    ),
    abortConfig = list(
      criteriaList = list(
        list(
          failureType = "FAILED"|"REJECTED"|"TIMED_OUT"|"ALL",
          action = "CANCEL",
          thresholdPercentage = 123.0,
          minNumberOfExecutedThings = 123
        )
      )
    ),
    timeoutConfig = list(
      inProgressTimeoutInMinutes = 123
    )
  ),
  deploymentPolicies = list(
    failureHandlingPolicy = "ROLLBACK"|"DO_NOTHING",
    componentUpdatePolicy = list(
      timeoutInSeconds = 123,
      action = "NOTIFY_COMPONENTS"|"SKIP_NOTIFY_COMPONENTS"
    ),
    configurationValidationPolicy = list(
      timeoutInSeconds = 123
    )
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
