% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_delete_component}
\alias{greengrassv2_delete_component}
\title{Deletes a version of a component from AWS IoT Greengrass}
\usage{
greengrassv2_delete_component(arn)
}
\arguments{
\item{arn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the component version.}
}
\description{
Deletes a version of a component from AWS IoT Greengrass.

This operation deletes the component's recipe and artifacts. As a
result, deployments that refer to this component version will fail. If
you have deployments that use this component version, you can remove the
component from the deployment or update the deployment to use a valid
version.
}
\section{Request syntax}{
\preformatted{svc$delete_component(
  arn = "string"
)
}
}

\keyword{internal}
