% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_delete_core_device}
\alias{greengrassv2_delete_core_device}
\title{Deletes a AWS IoT Greengrass core device, which is an AWS IoT thing}
\usage{
greengrassv2_delete_core_device(coreDeviceThingName)
}
\arguments{
\item{coreDeviceThingName}{[required] The name of the core device. This is also the name of the AWS IoT thing.}
}
\description{
Deletes a AWS IoT Greengrass core device, which is an AWS IoT thing.
This operation removes the core device from the list of core devices.
This operation doesn't delete the AWS IoT thing. For more information
about how to delete the AWS IoT thing, see
\href{https://docs.aws.amazon.com/iot/latest/apireference/API_DeleteThing.html}{DeleteThing}
in the \emph{AWS IoT API Reference}.
}
\section{Request syntax}{
\preformatted{svc$delete_core_device(
  coreDeviceThingName = "string"
)
}
}

\keyword{internal}
