% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_describe_component}
\alias{greengrassv2_describe_component}
\title{Retrieves metadata for a version of a component}
\usage{
greengrassv2_describe_component(arn)
}
\arguments{
\item{arn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the component version.}
}
\description{
Retrieves metadata for a version of a component.
}
\section{Request syntax}{
\preformatted{svc$describe_component(
  arn = "string"
)
}
}

\keyword{internal}
