% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_get_component}
\alias{greengrassv2_get_component}
\title{Gets the recipe for a version of a component}
\usage{
greengrassv2_get_component(recipeOutputFormat, arn)
}
\arguments{
\item{recipeOutputFormat}{The format of the recipe.}

\item{arn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the component version.}
}
\description{
Gets the recipe for a version of a component. Core devices can call this
operation to identify the artifacts and requirements to install a
component.
}
\section{Request syntax}{
\preformatted{svc$get_component(
  recipeOutputFormat = "JSON"|"YAML",
  arn = "string"
)
}
}

\keyword{internal}
