% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_get_component_version_artifact}
\alias{greengrassv2_get_component_version_artifact}
\title{Gets the pre-signed URL to a component artifact in an S3 bucket}
\usage{
greengrassv2_get_component_version_artifact(arn, artifactName)
}
\arguments{
\item{arn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the component version.}

\item{artifactName}{[required] The name of the artifact.}
}
\description{
Gets the pre-signed URL to a component artifact in an S3 bucket. Core
devices can call this operation to identify the URL that they can use to
download an artifact to install.
}
\section{Request syntax}{
\preformatted{svc$get_component_version_artifact(
  arn = "string",
  artifactName = "string"
)
}
}

\keyword{internal}
