% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_get_core_device}
\alias{greengrassv2_get_core_device}
\title{Retrieves metadata for a AWS IoT Greengrass core device}
\usage{
greengrassv2_get_core_device(coreDeviceThingName)
}
\arguments{
\item{coreDeviceThingName}{[required] The name of the core device. This is also the name of the AWS IoT thing.}
}
\description{
Retrieves metadata for a AWS IoT Greengrass core device.
}
\section{Request syntax}{
\preformatted{svc$get_core_device(
  coreDeviceThingName = "string"
)
}
}

\keyword{internal}
