% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_get_deployment}
\alias{greengrassv2_get_deployment}
\title{Gets a deployment}
\usage{
greengrassv2_get_deployment(deploymentId)
}
\arguments{
\item{deploymentId}{[required] The ID of the deployment.}
}
\description{
Gets a deployment. Deployments define the components that run on AWS IoT
Greengrass core devices.
}
\section{Request syntax}{
\preformatted{svc$get_deployment(
  deploymentId = "string"
)
}
}

\keyword{internal}
