% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_list_component_versions}
\alias{greengrassv2_list_component_versions}
\title{Retrieves a paginated list of all versions for a component}
\usage{
greengrassv2_list_component_versions(arn, maxResults, nextToken)
}
\arguments{
\item{arn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the component version.}

\item{maxResults}{The maximum number of results to be returned per paginated request.}

\item{nextToken}{The token to be used for the next set of paginated results.}
}
\description{
Retrieves a paginated list of all versions for a component.
}
\section{Request syntax}{
\preformatted{svc$list_component_versions(
  arn = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
