% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_list_components}
\alias{greengrassv2_list_components}
\title{Retrieves a paginated list of component summaries}
\usage{
greengrassv2_list_components(scope, maxResults, nextToken)
}
\arguments{
\item{scope}{The scope of the components to list.

Default: \code{PRIVATE}}

\item{maxResults}{The maximum number of results to be returned per paginated request.}

\item{nextToken}{The token to be used for the next set of paginated results.}
}
\description{
Retrieves a paginated list of component summaries. This list includes
components that you have permission to view.
}
\section{Request syntax}{
\preformatted{svc$list_components(
  scope = "PRIVATE"|"PUBLIC",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
