% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_list_core_devices}
\alias{greengrassv2_list_core_devices}
\title{Retrieves a paginated list of AWS IoT Greengrass core devices}
\usage{
greengrassv2_list_core_devices(thingGroupArn, status, maxResults,
  nextToken)
}
\arguments{
\item{thingGroupArn}{The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the AWS IoT thing group by which to filter. If you specify this
parameter, the list includes only core devices that are members of this
thing group.}

\item{status}{The core device status by which to filter. If you specify this
parameter, the list includes only core devices that have this status.
Choose one of the following options:
\itemize{
\item \code{HEALTHY} – The AWS IoT Greengrass Core software and all components
run on the core device without issue.
\item \code{UNHEALTHY} – The AWS IoT Greengrass Core software or a component is
in a failed state on the core device.
}}

\item{maxResults}{The maximum number of results to be returned per paginated request.}

\item{nextToken}{The token to be used for the next set of paginated results.}
}
\description{
Retrieves a paginated list of AWS IoT Greengrass core devices.
}
\section{Request syntax}{
\preformatted{svc$list_core_devices(
  thingGroupArn = "string",
  status = "HEALTHY"|"UNHEALTHY",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
