% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_list_effective_deployments}
\alias{greengrassv2_list_effective_deployments}
\title{Retrieves a paginated list of deployment jobs that AWS IoT Greengrass
sends to AWS IoT Greengrass core devices}
\usage{
greengrassv2_list_effective_deployments(coreDeviceThingName, maxResults,
  nextToken)
}
\arguments{
\item{coreDeviceThingName}{[required] The name of the core device. This is also the name of the AWS IoT thing.}

\item{maxResults}{The maximum number of results to be returned per paginated request.}

\item{nextToken}{The token to be used for the next set of paginated results.}
}
\description{
Retrieves a paginated list of deployment jobs that AWS IoT Greengrass
sends to AWS IoT Greengrass core devices.
}
\section{Request syntax}{
\preformatted{svc$list_effective_deployments(
  coreDeviceThingName = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
