% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrassv2_operations.R
\name{greengrassv2_tag_resource}
\alias{greengrassv2_tag_resource}
\title{Adds tags to an AWS IoT Greengrass resource}
\usage{
greengrassv2_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the resource to tag.}

\item{tags}{[required] A list of key-value pairs that contain metadata for the resource. For
more information, see \href{https://docs.aws.amazon.com/greengrass/v2/}{Tag your resources} in the \emph{AWS IoT
Greengrass V2 Developer Guide}.}
}
\description{
Adds tags to an AWS IoT Greengrass resource. If a tag already exists for
the resource, this operation updates the tag's value.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
