% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_cancel_contact}
\alias{groundstation_cancel_contact}
\title{Cancels a contact with a specified contact ID}
\usage{
groundstation_cancel_contact(contactId)
}
\arguments{
\item{contactId}{[required] UUID of a contact.}
}
\description{
Cancels a contact with a specified contact ID.
}
\section{Request syntax}{
\preformatted{svc$cancel_contact(
  contactId = "string"
)
}
}

\keyword{internal}
