% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_create_dataflow_endpoint_group}
\alias{groundstation_create_dataflow_endpoint_group}
\title{Creates a DataflowEndpoint group containing the specified list of
DataflowEndpoint objects}
\usage{
groundstation_create_dataflow_endpoint_group(endpointDetails, tags)
}
\arguments{
\item{endpointDetails}{[required] Endpoint details of each endpoint in the dataflow endpoint group.}

\item{tags}{Tags of a dataflow endpoint group.}
}
\description{
Creates a \code{DataflowEndpoint} group containing the specified list of
\code{DataflowEndpoint} objects.

The \code{name} field in each endpoint is used in your mission profile
\code{DataflowEndpointConfig} to specify which endpoints to use during a
contact.

When a contact uses multiple \code{DataflowEndpointConfig} objects, each
\code{Config} must match a \code{DataflowEndpoint} in the same group.
}
\section{Request syntax}{
\preformatted{svc$create_dataflow_endpoint_group(
  endpointDetails = list(
    list(
      endpoint = list(
        address = list(
          name = "string",
          port = 123
        ),
        mtu = 123,
        name = "string",
        status = "created"|"creating"|"deleted"|"deleting"|"failed"
      ),
      securityDetails = list(
        roleArn = "string",
        securityGroupIds = list(
          "string"
        ),
        subnetIds = list(
          "string"
        )
      )
    )
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
