% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_create_mission_profile}
\alias{groundstation_create_mission_profile}
\title{Creates a mission profile}
\usage{
groundstation_create_mission_profile(contactPostPassDurationSeconds,
  contactPrePassDurationSeconds, dataflowEdges,
  minimumViableContactDurationSeconds, name, tags, trackingConfigArn)
}
\arguments{
\item{contactPostPassDurationSeconds}{Amount of time after a contact ends that you'd like to receive a
CloudWatch event indicating the pass has finished.}

\item{contactPrePassDurationSeconds}{Amount of time prior to contact start you'd like to receive a CloudWatch
event indicating an upcoming pass.}

\item{dataflowEdges}{[required] A list of lists of ARNs. Each list of ARNs is an edge, with a from
\code{Config} and a to \code{Config}.}

\item{minimumViableContactDurationSeconds}{[required] Smallest amount of time in seconds that you'd like to see for an
available contact. AWS Ground Station will not present you with contacts
shorter than this duration.}

\item{name}{[required] Name of a mission profile.}

\item{tags}{Tags assigned to a mission profile.}

\item{trackingConfigArn}{[required] ARN of a tracking \code{Config}.}
}
\description{
Creates a mission profile.
}
\details{
\code{dataflowEdges} is a list of lists of strings. Each lower level list of
strings has two elements: a \emph{from ARN} and a \emph{to ARN}.
}
\section{Request syntax}{
\preformatted{svc$create_mission_profile(
  contactPostPassDurationSeconds = 123,
  contactPrePassDurationSeconds = 123,
  dataflowEdges = list(
    list(
      "string"
    )
  ),
  minimumViableContactDurationSeconds = 123,
  name = "string",
  tags = list(
    "string"
  ),
  trackingConfigArn = "string"
)
}
}

\keyword{internal}
