% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_delete_config}
\alias{groundstation_delete_config}
\title{Deletes a Config}
\usage{
groundstation_delete_config(configId, configType)
}
\arguments{
\item{configId}{[required] UUID of a \code{Config}.}

\item{configType}{[required] Type of a \code{Config}.}
}
\description{
Deletes a \code{Config}.
}
\section{Request syntax}{
\preformatted{svc$delete_config(
  configId = "string",
  configType = "antenna-downlink"|"antenna-downlink-demod-decode"|"antenna-uplink"|"dataflow-endpoint"|"tracking"|"uplink-echo"
)
}
}

\keyword{internal}
