% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_delete_mission_profile}
\alias{groundstation_delete_mission_profile}
\title{Deletes a mission profile}
\usage{
groundstation_delete_mission_profile(missionProfileId)
}
\arguments{
\item{missionProfileId}{[required] UUID of a mission profile.}
}
\description{
Deletes a mission profile.
}
\section{Request syntax}{
\preformatted{svc$delete_mission_profile(
  missionProfileId = "string"
)
}
}

\keyword{internal}
