% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_describe_contact}
\alias{groundstation_describe_contact}
\title{Describes an existing contact}
\usage{
groundstation_describe_contact(contactId)
}
\arguments{
\item{contactId}{[required] UUID of a contact.}
}
\description{
Describes an existing contact.
}
\section{Request syntax}{
\preformatted{svc$describe_contact(
  contactId = "string"
)
}
}

\keyword{internal}
