% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_get_config}
\alias{groundstation_get_config}
\title{Returns Config information}
\usage{
groundstation_get_config(configId, configType)
}
\arguments{
\item{configId}{[required] UUID of a \code{Config}.}

\item{configType}{[required] Type of a \code{Config}.}
}
\description{
Returns \code{Config} information.

Only one \code{Config} response can be returned.
}
\section{Request syntax}{
\preformatted{svc$get_config(
  configId = "string",
  configType = "antenna-downlink"|"antenna-downlink-demod-decode"|"antenna-uplink"|"dataflow-endpoint"|"tracking"|"uplink-echo"
)
}
}

\keyword{internal}
