% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_get_minute_usage}
\alias{groundstation_get_minute_usage}
\title{Returns the number of minutes used by account}
\usage{
groundstation_get_minute_usage(month, year)
}
\arguments{
\item{month}{[required] The month being requested, with a value of 1-12.}

\item{year}{[required] The year being requested, in the format of YYYY.}
}
\description{
Returns the number of minutes used by account.
}
\section{Request syntax}{
\preformatted{svc$get_minute_usage(
  month = 123,
  year = 123
)
}
}

\keyword{internal}
