% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_list_configs}
\alias{groundstation_list_configs}
\title{Returns a list of Config objects}
\usage{
groundstation_list_configs(maxResults, nextToken)
}
\arguments{
\item{maxResults}{Maximum number of \code{Configs} returned.}

\item{nextToken}{Next token returned in the request of a previous
\code{\link[=groundstation_list_configs]{list_configs}} call. Used to get the next
page of results.}
}
\description{
Returns a list of \code{Config} objects.
}
\section{Request syntax}{
\preformatted{svc$list_configs(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
