% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_list_contacts}
\alias{groundstation_list_contacts}
\title{Returns a list of contacts}
\usage{
groundstation_list_contacts(endTime, groundStation, maxResults,
  missionProfileArn, nextToken, satelliteArn, startTime, statusList)
}
\arguments{
\item{endTime}{[required] End time of a contact.}

\item{groundStation}{Name of a ground station.}

\item{maxResults}{Maximum number of contacts returned.}

\item{missionProfileArn}{ARN of a mission profile.}

\item{nextToken}{Next token returned in the request of a previous
\code{\link[=groundstation_list_contacts]{list_contacts}} call. Used to get the
next page of results.}

\item{satelliteArn}{ARN of a satellite.}

\item{startTime}{[required] Start time of a contact.}

\item{statusList}{[required] Status of a contact reservation.}
}
\description{
Returns a list of contacts.

If \code{statusList} contains AVAILABLE, the request must include
\code{groundStation}, \code{missionprofileArn}, and \code{satelliteArn}.
}
\section{Request syntax}{
\preformatted{svc$list_contacts(
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  groundStation = "string",
  maxResults = 123,
  missionProfileArn = "string",
  nextToken = "string",
  satelliteArn = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  statusList = list(
    "AVAILABLE"|"AWS_CANCELLED"|"AWS_FAILED"|"CANCELLED"|"CANCELLING"|"COMPLETED"|"FAILED"|"FAILED_TO_SCHEDULE"|"PASS"|"POSTPASS"|"PREPASS"|"SCHEDULED"|"SCHEDULING"
  )
)
}
}

\keyword{internal}
