% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_list_dataflow_endpoint_groups}
\alias{groundstation_list_dataflow_endpoint_groups}
\title{Returns a list of DataflowEndpoint groups}
\usage{
groundstation_list_dataflow_endpoint_groups(maxResults, nextToken)
}
\arguments{
\item{maxResults}{Maximum number of dataflow endpoint groups returned.}

\item{nextToken}{Next token returned in the request of a previous
\code{\link[=groundstation_list_dataflow_endpoint_groups]{list_dataflow_endpoint_groups}}
call. Used to get the next page of results.}
}
\description{
Returns a list of \code{DataflowEndpoint} groups.
}
\section{Request syntax}{
\preformatted{svc$list_dataflow_endpoint_groups(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
