% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_list_ground_stations}
\alias{groundstation_list_ground_stations}
\title{Returns a list of ground stations}
\usage{
groundstation_list_ground_stations(maxResults, nextToken)
}
\arguments{
\item{maxResults}{Maximum number of ground stations returned.}

\item{nextToken}{Next token that can be supplied in the next call to get the next page of
ground stations.}
}
\description{
Returns a list of ground stations.
}
\section{Request syntax}{
\preformatted{svc$list_ground_stations(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
