% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_list_mission_profiles}
\alias{groundstation_list_mission_profiles}
\title{Returns a list of mission profiles}
\usage{
groundstation_list_mission_profiles(maxResults, nextToken)
}
\arguments{
\item{maxResults}{Maximum number of mission profiles returned.}

\item{nextToken}{Next token returned in the request of a previous \code{ListMissionProfiles}
call. Used to get the next page of results.}
}
\description{
Returns a list of mission profiles.
}
\section{Request syntax}{
\preformatted{svc$list_mission_profiles(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
