% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_list_satellites}
\alias{groundstation_list_satellites}
\title{Returns a list of satellites}
\usage{
groundstation_list_satellites(maxResults, nextToken)
}
\arguments{
\item{maxResults}{Maximum number of satellites returned.}

\item{nextToken}{Next token that can be supplied in the next call to get the next page of
satellites.}
}
\description{
Returns a list of satellites.
}
\section{Request syntax}{
\preformatted{svc$list_satellites(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
