% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_list_tags_for_resource}
\alias{groundstation_list_tags_for_resource}
\title{Returns a list of tags or a specified resource}
\usage{
groundstation_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] ARN of a resource.}
}
\description{
Returns a list of tags or a specified resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
