% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_reserve_contact}
\alias{groundstation_reserve_contact}
\title{Reserves a contact using specified parameters}
\usage{
groundstation_reserve_contact(endTime, groundStation, missionProfileArn,
  satelliteArn, startTime, tags)
}
\arguments{
\item{endTime}{[required] End time of a contact.}

\item{groundStation}{[required] Name of a ground station.}

\item{missionProfileArn}{[required] ARN of a mission profile.}

\item{satelliteArn}{[required] ARN of a satellite}

\item{startTime}{[required] Start time of a contact.}

\item{tags}{Tags assigned to a contact.}
}
\description{
Reserves a contact using specified parameters.
}
\section{Request syntax}{
\preformatted{svc$reserve_contact(
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  groundStation = "string",
  missionProfileArn = "string",
  satelliteArn = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
