% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_tag_resource}
\alias{groundstation_tag_resource}
\title{Assigns a tag to a resource}
\usage{
groundstation_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] ARN of a resource tag.}

\item{tags}{Tags assigned to a resource.}
}
\description{
Assigns a tag to a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
