% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_untag_resource}
\alias{groundstation_untag_resource}
\title{Deassigns a resource tag}
\usage{
groundstation_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] ARN of a resource.}

\item{tagKeys}{[required] Keys of a resource tag.}
}
\description{
Deassigns a resource tag.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
