% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_update_mission_profile}
\alias{groundstation_update_mission_profile}
\title{Updates a mission profile}
\usage{
groundstation_update_mission_profile(contactPostPassDurationSeconds,
  contactPrePassDurationSeconds, dataflowEdges,
  minimumViableContactDurationSeconds, missionProfileId, name,
  trackingConfigArn)
}
\arguments{
\item{contactPostPassDurationSeconds}{Amount of time after a contact ends that you'd like to receive a
CloudWatch event indicating the pass has finished.}

\item{contactPrePassDurationSeconds}{Amount of time after a contact ends that you'd like to receive a
CloudWatch event indicating the pass has finished.}

\item{dataflowEdges}{A list of lists of ARNs. Each list of ARNs is an edge, with a \emph{from}
\code{Config} and a \emph{to} \code{Config}.}

\item{minimumViableContactDurationSeconds}{Smallest amount of time in seconds that you'd like to see for an
available contact. AWS Ground Station will not present you with contacts
shorter than this duration.}

\item{missionProfileId}{[required] UUID of a mission profile.}

\item{name}{Name of a mission profile.}

\item{trackingConfigArn}{ARN of a tracking \code{Config}.}
}
\description{
Updates a mission profile.
}
\details{
Updating a mission profile will not update the execution parameters for
existing future contacts.
}
\section{Request syntax}{
\preformatted{svc$update_mission_profile(
  contactPostPassDurationSeconds = 123,
  contactPrePassDurationSeconds = 123,
  dataflowEdges = list(
    list(
      "string"
    )
  ),
  minimumViableContactDurationSeconds = 123,
  missionProfileId = "string",
  name = "string",
  trackingConfigArn = "string"
)
}
}

\keyword{internal}
