% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_accept_invitation}
\alias{guardduty_accept_invitation}
\title{Accepts the invitation to be monitored by a master GuardDuty account}
\usage{
guardduty_accept_invitation(DetectorId, MasterId, InvitationId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty member account.}

\item{MasterId}{[required] The account ID of the master GuardDuty account whose invitation you\'re
accepting.}

\item{InvitationId}{[required] The value that is used to validate the master account to the member
account.}
}
\description{
Accepts the invitation to be monitored by a master GuardDuty account.
}
\section{Request syntax}{
\preformatted{svc$accept_invitation(
  DetectorId = "string",
  MasterId = "string",
  InvitationId = "string"
)
}
}

\keyword{internal}
