% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_archive_findings}
\alias{guardduty_archive_findings}
\title{Archives Amazon GuardDuty findings specified by the list of finding IDs}
\usage{
guardduty_archive_findings(DetectorId, FindingIds)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector that specifies the GuardDuty service whose
findings you want to archive.}

\item{FindingIds}{[required] IDs of the findings that you want to archive.}
}
\description{
Archives Amazon GuardDuty findings specified by the list of finding IDs.
}
\section{Request syntax}{
\preformatted{svc$archive_findings(
  DetectorId = "string",
  FindingIds = list(
    "string"
  )
)
}
}

\keyword{internal}
