% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_threat_intel_set}
\alias{guardduty_delete_threat_intel_set}
\title{Deletes ThreatIntelSet specified by the ThreatIntelSet ID}
\usage{
guardduty_delete_threat_intel_set(DetectorId, ThreatIntelSetId)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose ThreatIntelSet you want to delete.}

\item{ThreatIntelSetId}{[required] The unique ID that specifies the ThreatIntelSet that you want to delete.}
}
\description{
Deletes ThreatIntelSet specified by the ThreatIntelSet ID.
}
\section{Request syntax}{
\preformatted{guardduty$delete_threat_intel_set(
  DetectorId = "string",
  ThreatIntelSetId = "string"
)
}
}

\keyword{internal}
