% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_findings_feedback}
\alias{guardduty_update_findings_feedback}
\title{Marks specified Amazon GuardDuty findings as useful or not useful}
\usage{
guardduty_update_findings_feedback(DetectorId, FindingIds, Feedback,
  Comments)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector that specifies the GuardDuty service whose
findings you want to mark as useful or not useful.}

\item{FindingIds}{[required] IDs of the findings that you want to mark as useful or not useful.}

\item{Feedback}{[required] Valid values: USEFUL \| NOT\_USEFUL}

\item{Comments}{Additional feedback about the GuardDuty findings.}
}
\description{
Marks specified Amazon GuardDuty findings as useful or not useful.
}
\section{Request syntax}{
\preformatted{svc$update_findings_feedback(
  DetectorId = "string",
  FindingIds = list(
    "string"
  ),
  Feedback = "USEFUL"|"NOT_USEFUL",
  Comments = "string"
)
}
}

\keyword{internal}
