% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_publishing_destination}
\alias{guardduty_update_publishing_destination}
\title{Updates information about the publishing destination specified by the
destinationId}
\usage{
guardduty_update_publishing_destination(DetectorId, DestinationId,
  DestinationProperties)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector associated with the publishing destinations to
update.}

\item{DestinationId}{[required] The ID of the publishing destination to update.}

\item{DestinationProperties}{A \code{DestinationProperties} object that includes the \code{DestinationArn} and
\code{KmsKeyArn} of the publishing destination.}
}
\description{
Updates information about the publishing destination specified by the
\code{destinationId}.
}
\section{Request syntax}{
\preformatted{svc$update_publishing_destination(
  DetectorId = "string",
  DestinationId = "string",
  DestinationProperties = list(
    DestinationArn = "string",
    KmsKeyArn = "string"
  )
)
}
}

\keyword{internal}
