% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_to_csv.R
\name{h5_to_csv}
\alias{h5_to_csv}
\title{Batch convert SLEAP H5 files into DeepLabCut-like CSVs}
\usage{
h5_to_csv(h5_folder, save_folder = NA)
}
\arguments{
\item{h5_folder}{Path to the folder containing SLEAP H5s.}

\item{save_folder}{Path to the folder where DeepLabCut-like CSVs will be saved.}
}
\value{
Converted CSVs stored in the \code{save_folder} directory.
}
\description{
This function loads a directory containing H5 files generated from SLEAP.
It processes the first track in the CSV (assuming single-animal) and rearranges
the tracked points to match the style of that in DeepLabCut. Generated CSVs
are all saved to a separate directory, but can be saved in the H5 directory if
no \code{save_folder} parameter is assigned.
}
