% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_describe_events}
\alias{health_describe_events}
\title{Returns information about events that meet the specified filter criteria}
\usage{
health_describe_events(filter, nextToken, maxResults, locale)
}
\arguments{
\item{filter}{Values to narrow the results returned.}

\item{nextToken}{If the results of a search are large, only a portion of the results are
returned, and a \code{nextToken} pagination token is returned in the
response. To retrieve the next batch of results, reissue the search
request and include the returned token. When all results have been
returned, the response does not contain a pagination token value.}

\item{maxResults}{The maximum number of items to return in one batch, between 10 and 100,
inclusive.}

\item{locale}{The locale (language) to return information in. English (en) is the
default and the only supported value at this time.}
}
\description{
Returns information about events that meet the specified filter
criteria. Events are returned in a summary form and do not include the
detailed description, any additional metadata that depends on the event
type, or any affected resources. To retrieve that information, use the
\href{https://docs.aws.amazon.com/health/latest/APIReference/API_DescribeEventDetails.html}{DescribeEventDetails}
and
\href{https://docs.aws.amazon.com/health/latest/APIReference/API_DescribeAffectedEntities.html}{DescribeAffectedEntities}
operations.

If no filter criteria are specified, all events are returned. Results
are sorted by \code{lastModifiedTime}, starting with the most recent event.
\itemize{
\item When you call the \code{DescribeEvents} operation and specify an entity
for the \code{entityValues} parameter, AWS Health might return public
events that aren't specific to that resource. For example, if you
call \code{DescribeEvents} and specify an ID for an Amazon Elastic
Compute Cloud (Amazon EC2) instance, AWS Health might return events
that aren't specific to that resource or service. To get events that
are specific to a service, use the \code{services} parameter in the
\code{filter} object. For more information, see
\href{https://docs.aws.amazon.com/health/latest/APIReference/API_Event.html}{Event}.
\item This API operation uses pagination. Specify the \code{nextToken}
parameter in the next request to return more results.
}
}
\section{Request syntax}{
\preformatted{svc$describe_events(
  filter = list(
    eventArns = list(
      "string"
    ),
    eventTypeCodes = list(
      "string"
    ),
    services = list(
      "string"
    ),
    regions = list(
      "string"
    ),
    availabilityZones = list(
      "string"
    ),
    startTimes = list(
      list(
        from = as.POSIXct(
          "2015-01-01"
        ),
        to = as.POSIXct(
          "2015-01-01"
        )
      )
    ),
    endTimes = list(
      list(
        from = as.POSIXct(
          "2015-01-01"
        ),
        to = as.POSIXct(
          "2015-01-01"
        )
      )
    ),
    lastUpdatedTimes = list(
      list(
        from = as.POSIXct(
          "2015-01-01"
        ),
        to = as.POSIXct(
          "2015-01-01"
        )
      )
    ),
    entityArns = list(
      "string"
    ),
    entityValues = list(
      "string"
    ),
    eventTypeCategories = list(
      "issue"|"accountNotification"|"scheduledChange"|"investigation"
    ),
    tags = list(
      list(
        "string"
      )
    ),
    eventStatusCodes = list(
      "open"|"closed"|"upcoming"
    )
  ),
  nextToken = "string",
  maxResults = 123,
  locale = "string"
)
}
}

\keyword{internal}
