% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_enable_health_service_access_for_organization}
\alias{health_enable_health_service_access_for_organization}
\title{Calling this operation enables AWS Health to work with AWS Organizations}
\usage{
health_enable_health_service_access_for_organization()
}
\description{
Calling this operation enables AWS Health to work with AWS
Organizations. This applies a Service Linked Role (SLR) to the master
account in the organization. To learn more about the steps in this
process, visit enabling service access for AWS Health in AWS
Organizations. To call this operation, you must sign in as an IAM user,
assume an IAM role, or sign in as the root user (not recommended) in the
organization\'s master account.
}
\keyword{internal}
