% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_create_fhir_datastore}
\alias{healthlake_create_fhir_datastore}
\title{Creates a Data Store that can ingest and export FHIR formatted data}
\usage{
healthlake_create_fhir_datastore(DatastoreName, DatastoreTypeVersion,
  PreloadDataConfig, ClientToken)
}
\arguments{
\item{DatastoreName}{The user generated name for the Data Store.}

\item{DatastoreTypeVersion}{[required] The FHIR version of the Data Store. The only supported version is R4.}

\item{PreloadDataConfig}{Optional parameter to preload data upon creation of the Data Store.
Currently, the only supported preloaded data is synthetic data generated
from Synthea.}

\item{ClientToken}{Optional user provided token used for ensuring idempotency.}
}
\description{
Creates a Data Store that can ingest and export FHIR formatted data.
}
\section{Request syntax}{
\preformatted{svc$create_fhir_datastore(
  DatastoreName = "string",
  DatastoreTypeVersion = "R4",
  PreloadDataConfig = list(
    PreloadDataType = "SYNTHEA"
  ),
  ClientToken = "string"
)
}
}

\keyword{internal}
