% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_delete_fhir_datastore}
\alias{healthlake_delete_fhir_datastore}
\title{Deletes a Data Store}
\usage{
healthlake_delete_fhir_datastore(DatastoreId)
}
\arguments{
\item{DatastoreId}{The AWS-generated ID for the Data Store to be deleted.}
}
\description{
Deletes a Data Store.
}
\section{Request syntax}{
\preformatted{svc$delete_fhir_datastore(
  DatastoreId = "string"
)
}
}

\keyword{internal}
