% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_list_fhir_datastores}
\alias{healthlake_list_fhir_datastores}
\title{Lists all FHIR Data Stores that are in the user’s account, regardless of
Data Store status}
\usage{
healthlake_list_fhir_datastores(Filter, NextToken, MaxResults)
}
\arguments{
\item{Filter}{Lists all filters associated with a FHIR Data Store request.}

\item{NextToken}{Fetches the next page of Data Stores when results are paginated.}

\item{MaxResults}{The maximum number of Data Stores returned in a single page of a
ListFHIRDatastoresRequest call.}
}
\description{
Lists all FHIR Data Stores that are in the user’s account, regardless of
Data Store status.
}
\section{Request syntax}{
\preformatted{svc$list_fhir_datastores(
  Filter = list(
    DatastoreName = "string",
    DatastoreStatus = "CREATING"|"ACTIVE"|"DELETING"|"DELETED",
    CreatedBefore = as.POSIXct(
      "2015-01-01"
    ),
    CreatedAfter = as.POSIXct(
      "2015-01-01"
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
