% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_start_fhir_export_job}
\alias{healthlake_start_fhir_export_job}
\title{Begins a FHIR export job}
\usage{
healthlake_start_fhir_export_job(JobName, OutputDataConfig, DatastoreId,
  DataAccessRoleArn, ClientToken)
}
\arguments{
\item{JobName}{The user generated name for an export job.}

\item{OutputDataConfig}{[required] The output data configuration that was supplied when the export job was
created.}

\item{DatastoreId}{[required] The AWS generated ID for the Data Store from which files are being
exported for an export job.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name used during the initiation of the job.}

\item{ClientToken}{[required] An optional user provided token used for ensuring idempotency.}
}
\description{
Begins a FHIR export job.
}
\section{Request syntax}{
\preformatted{svc$start_fhir_export_job(
  JobName = "string",
  OutputDataConfig = list(
    S3Uri = "string"
  ),
  DatastoreId = "string",
  DataAccessRoleArn = "string",
  ClientToken = "string"
)
}
}

\keyword{internal}
