% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_start_fhir_import_job}
\alias{healthlake_start_fhir_import_job}
\title{Begins a FHIR Import job}
\usage{
healthlake_start_fhir_import_job(JobName, InputDataConfig, DatastoreId,
  DataAccessRoleArn, ClientToken)
}
\arguments{
\item{JobName}{The name of the FHIR Import job in the StartFHIRImport job request.}

\item{InputDataConfig}{[required] The input properties of the FHIR Import job in the StartFHIRImport job
request.}

\item{DatastoreId}{[required] The AWS-generated Data Store ID.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) that gives Amazon HealthLake access
permission.}

\item{ClientToken}{[required] Optional user provided token used for ensuring idempotency.}
}
\description{
Begins a FHIR Import job.
}
\section{Request syntax}{
\preformatted{svc$start_fhir_import_job(
  JobName = "string",
  InputDataConfig = list(
    S3Uri = "string"
  ),
  DatastoreId = "string",
  DataAccessRoleArn = "string",
  ClientToken = "string"
)
}
}

\keyword{internal}
