% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycode_operations.R
\name{honeycode_batch_delete_table_rows}
\alias{honeycode_batch_delete_table_rows}
\title{The BatchDeleteTableRows API allows you to delete one or more rows from
a table in a workbook}
\usage{
honeycode_batch_delete_table_rows(workbookId, tableId, rowIds,
  clientRequestToken)
}
\arguments{
\item{workbookId}{[required] The ID of the workbook where the rows are being deleted.

If a workbook with the specified id could not be found, this API throws
ResourceNotFoundException.}

\item{tableId}{[required] The ID of the table where the rows are being deleted.

If a table with the specified id could not be found, this API throws
ResourceNotFoundException.}

\item{rowIds}{[required] The list of row ids to delete from the table. You need to specify at
least one row id in this list.

Note that if one of the row ids provided in the request does not exist
in the table, then the request fails and no rows are deleted from the
table.}

\item{clientRequestToken}{The request token for performing the delete action. Request tokens help
to identify duplicate requests. If a call times out or fails due to a
transient error like a failed network connection, you can retry the call
with the same request token. The service ensures that if the first call
using that request token is successfully performed, the second call will
not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot
use request tokens to dedupe requests spanning hours or days.}
}
\description{
The BatchDeleteTableRows API allows you to delete one or more rows from
a table in a workbook. You need to specify the ids of the rows that you
want to delete from the table.
}
\section{Request syntax}{
\preformatted{svc$batch_delete_table_rows(
  workbookId = "string",
  tableId = "string",
  rowIds = list(
    "string"
  ),
  clientRequestToken = "string"
)
}
}

\keyword{internal}
