% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycode_operations.R
\name{honeycode_describe_table_data_import_job}
\alias{honeycode_describe_table_data_import_job}
\title{The DescribeTableDataImportJob API allows you to retrieve the status and
details of a table data import job}
\usage{
honeycode_describe_table_data_import_job(workbookId, tableId, jobId)
}
\arguments{
\item{workbookId}{[required] The ID of the workbook into which data was imported.

If a workbook with the specified id could not be found, this API throws
ResourceNotFoundException.}

\item{tableId}{[required] The ID of the table into which data was imported.

If a table with the specified id could not be found, this API throws
ResourceNotFoundException.}

\item{jobId}{[required] The ID of the job that was returned by the StartTableDataImportJob
request.

If a job with the specified id could not be found, this API throws
ResourceNotFoundException.}
}
\description{
The DescribeTableDataImportJob API allows you to retrieve the status and
details of a table data import job.
}
\section{Request syntax}{
\preformatted{svc$describe_table_data_import_job(
  workbookId = "string",
  tableId = "string",
  jobId = "string"
)
}
}

\keyword{internal}
