% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycode_operations.R
\name{honeycode_list_tables}
\alias{honeycode_list_tables}
\title{The ListTables API allows you to retrieve a list of all the tables in a
workbook}
\usage{
honeycode_list_tables(workbookId, maxResults, nextToken)
}
\arguments{
\item{workbookId}{[required] The ID of the workbook whose tables are being retrieved.

If a workbook with the specified id could not be found, this API throws
ResourceNotFoundException.}

\item{maxResults}{The maximum number of tables to return in each page of the results.}

\item{nextToken}{This parameter is optional. If a nextToken is not specified, the API
returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was
returned more than an hour back, the API will throw ValidationException.}
}
\description{
The ListTables API allows you to retrieve a list of all the tables in a
workbook.
}
\section{Request syntax}{
\preformatted{svc$list_tables(
  workbookId = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
