% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycode_operations.R
\name{honeycode_start_table_data_import_job}
\alias{honeycode_start_table_data_import_job}
\title{The StartTableDataImportJob API allows you to start an import job on a
table}
\usage{
honeycode_start_table_data_import_job(workbookId, dataSource,
  dataFormat, destinationTableId, importOptions, clientRequestToken)
}
\arguments{
\item{workbookId}{[required] The ID of the workbook where the rows are being imported.

If a workbook with the specified id could not be found, this API throws
ResourceNotFoundException.}

\item{dataSource}{[required] The source of the data that is being imported. The size of source must
be no larger than 100 MB. Source must have no more than 100,000 cells
and no more than 1,000 rows.}

\item{dataFormat}{[required] The format of the data that is being imported. Currently the only option
supported is "DELIMITED\\_TEXT".}

\item{destinationTableId}{[required] The ID of the table where the rows are being imported.

If a table with the specified id could not be found, this API throws
ResourceNotFoundException.}

\item{importOptions}{[required] The options for customizing this import request.}

\item{clientRequestToken}{[required] The request token for performing the update action. Request tokens help
to identify duplicate requests. If a call times out or fails due to a
transient error like a failed network connection, you can retry the call
with the same request token. The service ensures that if the first call
using that request token is successfully performed, the second call will
not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot
use request tokens to dedupe requests spanning hours or days.}
}
\description{
The StartTableDataImportJob API allows you to start an import job on a
table. This API will only return the id of the job that was started. To
find out the status of the import request, you need to call the
DescribeTableDataImportJob API.
}
\section{Request syntax}{
\preformatted{svc$start_table_data_import_job(
  workbookId = "string",
  dataSource = list(
    dataSourceConfig = list(
      dataSourceUrl = "string"
    )
  ),
  dataFormat = "DELIMITED_TEXT",
  destinationTableId = "string",
  importOptions = list(
    destinationOptions = list(
      columnMap = list(
        list(
          columnIndex = 123
        )
      )
    ),
    delimitedTextOptions = list(
      delimiter = "string",
      hasHeaderRow = TRUE|FALSE,
      ignoreEmptyRows = TRUE|FALSE,
      dataCharacterEncoding = "UTF-8"|"US-ASCII"|"ISO-8859-1"|"UTF-16BE"|"UTF-16LE"|"UTF-16"
    )
  ),
  clientRequestToken = "string"
)
}
}

\keyword{internal}
