% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_create_instance_profile}
\alias{iam_create_instance_profile}
\title{Creates a new instance profile}
\usage{
iam_create_instance_profile(InstanceProfileName, Path)
}
\arguments{
\item{InstanceProfileName}{[required] The name of the instance profile to create.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: \_+=,.@-}

\item{Path}{The path to the instance profile. For more information about paths, see
\href{https://docs.aws.amazon.com/IAM/latest/UserGuide/Using_Identifiers.html}{IAM Identifiers}
in the \emph{IAM User Guide}.

This parameter is optional. If it is not included, it defaults to a
slash (/).

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters
consisting of either a forward slash (/) by itself or a string that must
begin and end with forward slashes. In addition, it can contain any
ASCII character from the ! (\code{U+0021}) through the DEL character
(\code{U+007F}), including most punctuation characters, digits, and upper and
lowercased letters.}
}
\description{
Creates a new instance profile. For information about instance profiles,
go to \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/AboutInstanceProfiles.html}{About Instance Profiles}.
}
\details{
For information about the number of instance profiles you can create,
see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/LimitationsOnEntities.html}{Limitations on IAM Entities}
in the \emph{IAM User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_instance_profile(
  InstanceProfileName = "string",
  Path = "string"
)
}
}

\examples{
# The following command creates an instance profile named Webserver that
# is ready to have a role attached and then be associated with an EC2
# instance.
\donttest{svc$create_instance_profile(
  InstanceProfileName = "Webserver"
)}

}
\keyword{internal}
