% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_policy_version}
\alias{iam_delete_policy_version}
\title{Deletes the specified version from the specified managed policy}
\usage{
iam_delete_policy_version(PolicyArn, VersionId)
}
\arguments{
\item{PolicyArn}{[required] The Amazon Resource Name (ARN) of the IAM policy from which you want to
delete a version.

For more information about ARNs, see \href{http://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resource Names (ARNs) and AWS Service Namespaces}
in the \emph{AWS General Reference}.}

\item{VersionId}{[required] The policy version to delete.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters that
consists of the lowercase letter 'v' followed by one or two digits,
and optionally followed by a period '.' and a string of letters and
digits.

For more information about managed policy versions, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/policies-managed-versions.html}{Versioning for Managed Policies}
in the \emph{IAM User Guide}.}
}
\description{
Deletes the specified version from the specified managed policy.
}
\details{
You cannot delete the default version from a policy using this API. To
delete the default version from a policy, use DeletePolicy. To find out
which version of a policy is marked as the default version, use
ListPolicyVersions.

For information about versions for managed policies, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/policies-managed-versions.html}{Versioning for Managed Policies}
in the \emph{IAM User Guide}.
}
\section{Request syntax}{
\preformatted{iam$delete_policy_version(
  PolicyArn = "string",
  VersionId = "string"
)
}
}

\keyword{internal}
