% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_account_password_policy}
\alias{iam_get_account_password_policy}
\title{Retrieves the password policy for the AWS account}
\usage{
iam_get_account_password_policy()
}
\description{
Retrieves the password policy for the AWS account. For more information
about using a password policy, go to \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/Using_ManagingPasswordPolicies.html}{Managing an IAM Password Policy}.
}
\section{Examples}{
\preformatted{# The following command displays details about the password policy for the
# current AWS account.
svc$get_account_password_policy()
}
}

\keyword{internal}
