% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_account_summary}
\alias{iam_get_account_summary}
\title{Retrieves information about IAM entity usage and IAM quotas in the AWS
account}
\usage{
iam_get_account_summary()
}
\description{
Retrieves information about IAM entity usage and IAM quotas in the AWS
account.
}
\details{
For information about limitations on IAM entities, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/LimitationsOnEntities.html}{Limitations on IAM Entities}
in the \emph{IAM User Guide}.
}
\examples{
\dontrun{
# The following command returns information about the IAM entity quotas
# and usage in the current AWS account.
svc$get_account_summary()
}

}
\keyword{internal}
