% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_list_groups}
\alias{identitystore_list_groups}
\title{Lists the attribute name and value of the group that you specified in
the search}
\usage{
identitystore_list_groups(IdentityStoreId, MaxResults, NextToken,
  Filters)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store, such as
d-1234567890. In this example, d- is a fixed prefix, and 1234567890 is a
randomly generated string which contains number and lower case letters.
This value is generated at the time that a new identity store is
created.}

\item{MaxResults}{The maximum number of results to be returned per request, which is used
in the \code{ListUsers} and \code{ListGroups} request to specify how many results
to return in one page. The length limit is 50 characters.}

\item{NextToken}{The pagination token used for the \code{ListUsers} and \code{ListGroups} APIs.
This value is generated by the identity store service and is returned in
the API response if the total results are more than the size of one
page, and when this token is used in the API request to search for the
next page.}

\item{Filters}{A list of \code{Filter} objects, which is used in the \code{ListUsers} and
\code{ListGroups} request.}
}
\description{
Lists the attribute name and value of the group that you specified in
the search. We only support \code{DisplayName} as a valid filter attribute
path currently, and filter is required. This API returns minimum
attributes, including \code{GroupId} and group \code{DisplayName} in the response.
}
\section{Request syntax}{
\preformatted{svc$list_groups(
  IdentityStoreId = "string",
  MaxResults = 123,
  NextToken = "string",
  Filters = list(
    list(
      AttributePath = "string",
      AttributeValue = "string"
    )
  )
)
}
}

\keyword{internal}
