% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_cancel_image_creation}
\alias{imagebuilder_cancel_image_creation}
\title{CancelImageCreation cancels the creation of Image}
\usage{
imagebuilder_cancel_image_creation(imageBuildVersionArn, clientToken)
}
\arguments{
\item{imageBuildVersionArn}{[required] The Amazon Resource Name (ARN) of the image whose creation you wish to
cancel.}

\item{clientToken}{[required] The idempotency token used to make this request idempotent.}
}
\description{
CancelImageCreation cancels the creation of Image. This operation may
only be used on images in a non-terminal state.
}
\section{Request syntax}{
\preformatted{svc$cancel_image_creation(
  imageBuildVersionArn = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
