% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_create_component}
\alias{imagebuilder_create_component}
\title{Creates a new component that can be used to build, validate, test and
assess your image}
\usage{
imagebuilder_create_component(name, semanticVersion, description,
  changeDescription, platform, data, uri, kmsKeyId, tags, clientToken)
}
\arguments{
\item{name}{[required] The name of the component.}

\item{semanticVersion}{[required] The semantic version of the component. This version to follow the
semantic version syntax. i.e. major.minor.patch. This could be versioned
like software 2.0.1 or date like 2019.12.01.}

\item{description}{CThe description of the component. Describes the contents of the
component.}

\item{changeDescription}{CThe change description of the component. Describes what change has been
made in this version. In other words what makes this version different
from other versions of this component.}

\item{platform}{[required] CThe platform of the component.}

\item{data}{CThe data of the component.}

\item{uri}{CThe uri of the component.}

\item{kmsKeyId}{The ID of the KMS key that should be used to encrypt this component.}

\item{tags}{CThe tags of the component.}

\item{clientToken}{[required] CThe idempotency token of the component.}
}
\description{
Creates a new component that can be used to build, validate, test and
assess your image.
}
\section{Request syntax}{
\preformatted{svc$create_component(
  name = "string",
  semanticVersion = "string",
  description = "string",
  changeDescription = "string",
  platform = "Windows"|"Linux",
  data = "string",
  uri = "string",
  kmsKeyId = "string",
  tags = list(
    "string"
  ),
  clientToken = "string"
)
}
}

\keyword{internal}
