% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_create_distribution_configuration}
\alias{imagebuilder_create_distribution_configuration}
\title{Creates a new distribution configuration}
\usage{
imagebuilder_create_distribution_configuration(name, description,
  distributions, tags, clientToken)
}
\arguments{
\item{name}{[required] The name of the distribution configuration.}

\item{description}{The description of the distribution configuration.}

\item{distributions}{[required] The distributions of the distribution configuration.}

\item{tags}{The tags of the distribution configuration.}

\item{clientToken}{[required] The idempotency token of the distribution configuration.}
}
\description{
Creates a new distribution configuration. Distribution configurations
define and configure the outputs of your pipeline.
}
\section{Request syntax}{
\preformatted{svc$create_distribution_configuration(
  name = "string",
  description = "string",
  distributions = list(
    list(
      region = "string",
      amiDistributionConfiguration = list(
        name = "string",
        description = "string",
        amiTags = list(
          "string"
        ),
        launchPermission = list(
          userIds = list(
            "string"
          ),
          userGroups = list(
            "string"
          )
        )
      ),
      licenseConfigurationArns = list(
        "string"
      )
    )
  ),
  tags = list(
    "string"
  ),
  clientToken = "string"
)
}
}

\keyword{internal}
