% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_create_image}
\alias{imagebuilder_create_image}
\title{Creates a new image}
\usage{
imagebuilder_create_image(imageRecipeArn, distributionConfigurationArn,
  infrastructureConfigurationArn, imageTestsConfiguration, tags,
  clientToken)
}
\arguments{
\item{imageRecipeArn}{[required] The Amazon Resource Name (ARN) of the image recipe that defines how
images are configured, tested and assessed.}

\item{distributionConfigurationArn}{The Amazon Resource Name (ARN) of the distribution configuration that
defines and configures the outputs of your pipeline.}

\item{infrastructureConfigurationArn}{[required] The Amazon Resource Name (ARN) of the infrastructure configuration that
defines the environment in which your image will be built and tested.}

\item{imageTestsConfiguration}{The image tests configuration of the image.}

\item{tags}{The tags of the image.}

\item{clientToken}{[required] The idempotency token used to make this request idempotent.}
}
\description{
Creates a new image. This request will create a new image along with all
of the configured output resources defined in the distribution
configuration.
}
\section{Request syntax}{
\preformatted{svc$create_image(
  imageRecipeArn = "string",
  distributionConfigurationArn = "string",
  infrastructureConfigurationArn = "string",
  imageTestsConfiguration = list(
    imageTestsEnabled = TRUE|FALSE,
    timeoutMinutes = 123
  ),
  tags = list(
    "string"
  ),
  clientToken = "string"
)
}
}

\keyword{internal}
