% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_create_image_recipe}
\alias{imagebuilder_create_image_recipe}
\title{Creates a new image recipe}
\usage{
imagebuilder_create_image_recipe(name, description, semanticVersion,
  components, parentImage, blockDeviceMappings, tags, clientToken)
}
\arguments{
\item{name}{[required] The name of the image recipe.}

\item{description}{The description of the image recipe.}

\item{semanticVersion}{[required] The semantic version of the image recipe.}

\item{components}{[required] The components of the image recipe.}

\item{parentImage}{[required] The parent image of the image recipe.}

\item{blockDeviceMappings}{The block device mappings of the image recipe.}

\item{tags}{The tags of the image recipe.}

\item{clientToken}{[required] The idempotency token used to make this request idempotent.}
}
\description{
Creates a new image recipe. Image Recipes defines how images are
configured, tested and assessed.
}
\section{Request syntax}{
\preformatted{svc$create_image_recipe(
  name = "string",
  description = "string",
  semanticVersion = "string",
  components = list(
    list(
      componentArn = "string"
    )
  ),
  parentImage = "string",
  blockDeviceMappings = list(
    list(
      deviceName = "string",
      ebs = list(
        encrypted = TRUE|FALSE,
        deleteOnTermination = TRUE|FALSE,
        iops = 123,
        kmsKeyId = "string",
        snapshotId = "string",
        volumeSize = 123,
        volumeType = "standard"|"io1"|"gp2"|"sc1"|"st1"
      ),
      virtualName = "string",
      noDevice = "string"
    )
  ),
  tags = list(
    "string"
  ),
  clientToken = "string"
)
}
}

\keyword{internal}
