% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_create_infrastructure_configuration}
\alias{imagebuilder_create_infrastructure_configuration}
\title{Creates a new infrastructure configuration}
\usage{
imagebuilder_create_infrastructure_configuration(name, description,
  instanceTypes, instanceProfileName, securityGroupIds, subnetId, logging,
  keyPair, terminateInstanceOnFailure, snsTopicArn, resourceTags, tags,
  clientToken)
}
\arguments{
\item{name}{[required] The name of the infrastructure configuration.}

\item{description}{The description of the infrastructure configuration.}

\item{instanceTypes}{The instance types of the infrastructure configuration. You can specify
one or more instance types to use for this build. The service will pick
one of these instance types based on availability.}

\item{instanceProfileName}{[required] The instance profile to associate with the instance used to customize
your EC2 AMI.}

\item{securityGroupIds}{The security group IDs to associate with the instance used to customize
your EC2 AMI.}

\item{subnetId}{The subnet ID in which to place the instance used to customize your EC2
AMI.}

\item{logging}{The logging configuration of the infrastructure configuration.}

\item{keyPair}{The key pair of the infrastructure configuration. This can be used to
log on to and debug the instance used to create your image.}

\item{terminateInstanceOnFailure}{The terminate instance on failure setting of the infrastructure
configuration. Set to false if you want Image Builder to retain the
instance used to configure your AMI if the build or test phase of your
workflow fails.}

\item{snsTopicArn}{The SNS topic on which to send image build events.}

\item{resourceTags}{The tags attached to the resource created by Image Builder.}

\item{tags}{The tags of the infrastructure configuration.}

\item{clientToken}{[required] The idempotency token used to make this request idempotent.}
}
\description{
Creates a new infrastructure configuration. An infrastructure
configuration defines the environment in which your image will be built
and tested.
}
\section{Request syntax}{
\preformatted{svc$create_infrastructure_configuration(
  name = "string",
  description = "string",
  instanceTypes = list(
    "string"
  ),
  instanceProfileName = "string",
  securityGroupIds = list(
    "string"
  ),
  subnetId = "string",
  logging = list(
    s3Logs = list(
      s3BucketName = "string",
      s3KeyPrefix = "string"
    )
  ),
  keyPair = "string",
  terminateInstanceOnFailure = TRUE|FALSE,
  snsTopicArn = "string",
  resourceTags = list(
    "string"
  ),
  tags = list(
    "string"
  ),
  clientToken = "string"
)
}
}

\keyword{internal}
