% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_delete_container_recipe}
\alias{imagebuilder_delete_container_recipe}
\title{Deletes a container recipe}
\usage{
imagebuilder_delete_container_recipe(containerRecipeArn)
}
\arguments{
\item{containerRecipeArn}{[required] The Amazon Resource Name (ARN) of the container recipe to delete.}
}
\description{
Deletes a container recipe.
}
\section{Request syntax}{
\preformatted{svc$delete_container_recipe(
  containerRecipeArn = "string"
)
}
}

\keyword{internal}
