% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_delete_distribution_configuration}
\alias{imagebuilder_delete_distribution_configuration}
\title{Deletes a distribution configuration}
\usage{
imagebuilder_delete_distribution_configuration(
  distributionConfigurationArn)
}
\arguments{
\item{distributionConfigurationArn}{[required] The Amazon Resource Name (ARN) of the distribution configuration to
delete.}
}
\description{
Deletes a distribution configuration.
}
\section{Request syntax}{
\preformatted{svc$delete_distribution_configuration(
  distributionConfigurationArn = "string"
)
}
}

\keyword{internal}
