% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_delete_image}
\alias{imagebuilder_delete_image}
\title{Deletes an image}
\usage{
imagebuilder_delete_image(imageBuildVersionArn)
}
\arguments{
\item{imageBuildVersionArn}{[required] The Amazon Resource Name (ARN) of the image to delete.}
}
\description{
Deletes an image.
}
\section{Request syntax}{
\preformatted{svc$delete_image(
  imageBuildVersionArn = "string"
)
}
}

\keyword{internal}
