% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_component}
\alias{imagebuilder_get_component}
\title{Gets a component object}
\usage{
imagebuilder_get_component(componentBuildVersionArn)
}
\arguments{
\item{componentBuildVersionArn}{[required] The Amazon Resource Name (ARN) of the component that you want to
retrieve. Regex requires "/\verb{\\\\d}+$" suffix.}
}
\description{
Gets a component object.
}
\section{Request syntax}{
\preformatted{svc$get_component(
  componentBuildVersionArn = "string"
)
}
}

\keyword{internal}
