% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_component_policy}
\alias{imagebuilder_get_component_policy}
\title{Gets a component policy}
\usage{
imagebuilder_get_component_policy(componentArn)
}
\arguments{
\item{componentArn}{[required] The Amazon Resource Name (ARN) of the component whose policy you want to
retrieve.}
}
\description{
Gets a component policy.
}
\section{Request syntax}{
\preformatted{svc$get_component_policy(
  componentArn = "string"
)
}
}

\keyword{internal}
