% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_container_recipe}
\alias{imagebuilder_get_container_recipe}
\title{Retrieves a container recipe}
\usage{
imagebuilder_get_container_recipe(containerRecipeArn)
}
\arguments{
\item{containerRecipeArn}{[required] The Amazon Resource Name (ARN) of the container recipe to retrieve.}
}
\description{
Retrieves a container recipe.
}
\section{Request syntax}{
\preformatted{svc$get_container_recipe(
  containerRecipeArn = "string"
)
}
}

\keyword{internal}
