% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_image}
\alias{imagebuilder_get_image}
\title{Gets an image}
\usage{
imagebuilder_get_image(imageBuildVersionArn)
}
\arguments{
\item{imageBuildVersionArn}{[required] The Amazon Resource Name (ARN) of the image that you wish to retrieve.}
}
\description{
Gets an image.
}
\section{Request syntax}{
\preformatted{svc$get_image(
  imageBuildVersionArn = "string"
)
}
}

\keyword{internal}
