% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_image_policy}
\alias{imagebuilder_get_image_policy}
\title{Gets an image policy}
\usage{
imagebuilder_get_image_policy(imageArn)
}
\arguments{
\item{imageArn}{[required] The Amazon Resource Name (ARN) of the image whose policy you wish to
retrieve.}
}
\description{
Gets an image policy.
}
\section{Request syntax}{
\preformatted{svc$get_image_policy(
  imageArn = "string"
)
}
}

\keyword{internal}
