% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_import_component}
\alias{imagebuilder_import_component}
\title{Imports a component and transforms its data into a component document}
\usage{
imagebuilder_import_component(name, semanticVersion, description,
  changeDescription, type, format, platform, data, uri, kmsKeyId, tags,
  clientToken)
}
\arguments{
\item{name}{[required] The name of the component.}

\item{semanticVersion}{[required] The semantic version of the component. This version to follow the
semantic version syntax. i.e. major.minor.patch. This could be versioned
like software 2.0.1 or date like 2019.12.01.}

\item{description}{The description of the component. Describes the contents of the
component.}

\item{changeDescription}{The change description of the component. Describes what change has been
made in this version. In other words what makes this version different
from other versions of this component.}

\item{type}{[required] The type of the component denotes whether the component is used to build
the image or only to test it.}

\item{format}{[required] The format of the resource that you wish to import as a component.}

\item{platform}{[required] The platform of the component.}

\item{data}{The data of the component.}

\item{uri}{The uri of the component.}

\item{kmsKeyId}{The ID of the KMS key that should be used to encrypt this component.}

\item{tags}{The tags of the component.}

\item{clientToken}{[required] The idempotency token of the component.}
}
\description{
Imports a component and transforms its data into a component document.
}
\section{Request syntax}{
\preformatted{svc$import_component(
  name = "string",
  semanticVersion = "string",
  description = "string",
  changeDescription = "string",
  type = "BUILD"|"TEST",
  format = "SHELL",
  platform = "Windows"|"Linux",
  data = "string",
  uri = "string",
  kmsKeyId = "string",
  tags = list(
    "string"
  ),
  clientToken = "string"
)
}
}

\keyword{internal}
