% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_components}
\alias{imagebuilder_list_components}
\title{Returns the list of component build versions for the specified semantic
version}
\usage{
imagebuilder_list_components(owner, filters, maxResults, nextToken)
}
\arguments{
\item{owner}{The owner defines whose components you wish to list. By default this
request will only show components owned by your account. You may use
this field to specify if you wish to view components owned by yourself,
Amazon, or those components that have been shared with you by other
customers.}

\item{filters}{}

\item{maxResults}{The maximum items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the NextToken from
a previously truncated response.}
}
\description{
Returns the list of component build versions for the specified semantic
version.
}
\section{Request syntax}{
\preformatted{svc$list_components(
  owner = "Self"|"Shared"|"Amazon",
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
