% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_container_recipes}
\alias{imagebuilder_list_container_recipes}
\title{Returns a list of container recipes}
\usage{
imagebuilder_list_container_recipes(owner, filters, maxResults,
  nextToken)
}
\arguments{
\item{owner}{Returns container recipes belonging to the specified owner, that have
been shared with you. You can omit this field to return container
recipes belonging to your account.}

\item{filters}{Request filters that are used to narrow the list of container images
that are returned.}

\item{maxResults}{The maximum number of results to return in the list.}

\item{nextToken}{Provides a token for pagination, which determines where to begin the
next set of results when the current set reaches the maximum for one
request.}
}
\description{
Returns a list of container recipes.
}
\section{Request syntax}{
\preformatted{svc$list_container_recipes(
  owner = "Self"|"Shared"|"Amazon",
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
