% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_distribution_configurations}
\alias{imagebuilder_list_distribution_configurations}
\title{Returns a list of distribution configurations}
\usage{
imagebuilder_list_distribution_configurations(filters, maxResults,
  nextToken)
}
\arguments{
\item{filters}{The filters.
\itemize{
\item \code{name} - The name of this distribution configuration.
}}

\item{maxResults}{The maximum items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the NextToken from
a previously truncated response.}
}
\description{
Returns a list of distribution configurations.
}
\section{Request syntax}{
\preformatted{svc$list_distribution_configurations(
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
