% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_pipelines}
\alias{imagebuilder_list_image_pipelines}
\title{Returns a list of image pipelines}
\usage{
imagebuilder_list_image_pipelines(filters, maxResults, nextToken)
}
\arguments{
\item{filters}{The filters.}

\item{maxResults}{The maximum items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the NextToken from
a previously truncated response.}
}
\description{
Returns a list of image pipelines.
}
\section{Request syntax}{
\preformatted{svc$list_image_pipelines(
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
