% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_recipes}
\alias{imagebuilder_list_image_recipes}
\title{Returns a list of image recipes}
\usage{
imagebuilder_list_image_recipes(owner, filters, maxResults, nextToken)
}
\arguments{
\item{owner}{The owner defines whose image recipes you wish to list. By default this
request will only show image recipes owned by your account. You may use
this field to specify if you wish to view image recipes owned by
yourself, Amazon, or those image recipes that have been shared with you
by other customers.}

\item{filters}{}

\item{maxResults}{The maximum items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the NextToken from
a previously truncated response.}
}
\description{
Returns a list of image recipes.
}
\section{Request syntax}{
\preformatted{svc$list_image_recipes(
  owner = "Self"|"Shared"|"Amazon",
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
