% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_infrastructure_configurations}
\alias{imagebuilder_list_infrastructure_configurations}
\title{Returns a list of infrastructure configurations}
\usage{
imagebuilder_list_infrastructure_configurations(filters, maxResults,
  nextToken)
}
\arguments{
\item{filters}{}

\item{maxResults}{The maximum items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the NextToken from
a previously truncated response.}
}
\description{
Returns a list of infrastructure configurations.
}
\section{Request syntax}{
\preformatted{svc$list_infrastructure_configurations(
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
