% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_put_component_policy}
\alias{imagebuilder_put_component_policy}
\title{Applies a policy to a component}
\usage{
imagebuilder_put_component_policy(componentArn, policy)
}
\arguments{
\item{componentArn}{[required] The Amazon Resource Name (ARN) of the component that this policy should
be applied to.}

\item{policy}{[required] The policy to apply.}
}
\description{
Applies a policy to a component.
}
\section{Request syntax}{
\preformatted{svc$put_component_policy(
  componentArn = "string",
  policy = "string"
)
}
}

\keyword{internal}
