% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_put_container_recipe_policy}
\alias{imagebuilder_put_container_recipe_policy}
\title{Applies a policy to a container image}
\usage{
imagebuilder_put_container_recipe_policy(containerRecipeArn, policy)
}
\arguments{
\item{containerRecipeArn}{[required] The Amazon Resource Name (ARN) of the container recipe that this policy
should be applied to.}

\item{policy}{[required] The policy to apply to the container recipe.}
}
\description{
Applies a policy to a container image. We recommend that you call the
RAM API CreateResourceShare
(https://docs.aws.amazon.com/ram/latest/APIReference/API\\_CreateResourceShare.html)
to share resources. If you call the Image Builder API
\code{PutContainerImagePolicy}, you must also call the RAM API
PromoteResourceShareCreatedFromPolicy
(https://docs.aws.amazon.com/ram/latest/APIReference/API\\_PromoteResourceShareCreatedFromPolicy.html)
in order for the resource to be visible to all principals with whom the
resource is shared.
}
\section{Request syntax}{
\preformatted{svc$put_container_recipe_policy(
  containerRecipeArn = "string",
  policy = "string"
)
}
}

\keyword{internal}
