% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_put_image_policy}
\alias{imagebuilder_put_image_policy}
\title{Applies a policy to an image}
\usage{
imagebuilder_put_image_policy(imageArn, policy)
}
\arguments{
\item{imageArn}{[required] The Amazon Resource Name (ARN) of the image that this policy should be
applied to.}

\item{policy}{[required] The policy to apply.}
}
\description{
Applies a policy to an image.
}
\section{Request syntax}{
\preformatted{svc$put_image_policy(
  imageArn = "string",
  policy = "string"
)
}
}

\keyword{internal}
