% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_put_image_recipe_policy}
\alias{imagebuilder_put_image_recipe_policy}
\title{Applies a policy to an image recipe}
\usage{
imagebuilder_put_image_recipe_policy(imageRecipeArn, policy)
}
\arguments{
\item{imageRecipeArn}{[required] The Amazon Resource Name (ARN) of the image recipe that this policy
should be applied to.}

\item{policy}{[required] The policy to apply.}
}
\description{
Applies a policy to an image recipe.
}
\section{Request syntax}{
\preformatted{svc$put_image_recipe_policy(
  imageRecipeArn = "string",
  policy = "string"
)
}
}

\keyword{internal}
