% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_update_distribution_configuration}
\alias{imagebuilder_update_distribution_configuration}
\title{Updates a new distribution configuration}
\usage{
imagebuilder_update_distribution_configuration(
  distributionConfigurationArn, description, distributions, clientToken)
}
\arguments{
\item{distributionConfigurationArn}{[required] The Amazon Resource Name (ARN) of the distribution configuration that
you wish to update.}

\item{description}{The description of the distribution configuration.}

\item{distributions}{The distributions of the distribution configuration.}

\item{clientToken}{[required] The idempotency token of the distribution configuration.}
}
\description{
Updates a new distribution configuration. Distribution configurations
define and configure the outputs of your pipeline.
}
\section{Request syntax}{
\preformatted{svc$update_distribution_configuration(
  distributionConfigurationArn = "string",
  description = "string",
  distributions = list(
    list(
      region = "string",
      amiDistributionConfiguration = list(
        name = "string",
        description = "string",
        amiTags = list(
          "string"
        ),
        launchPermission = list(
          userIds = list(
            "string"
          ),
          userGroups = list(
            "string"
          )
        )
      ),
      licenseConfigurationArns = list(
        "string"
      )
    )
  ),
  clientToken = "string"
)
}
}

\keyword{internal}
