% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_update_image_pipeline}
\alias{imagebuilder_update_image_pipeline}
\title{Updates a new image pipeline}
\usage{
imagebuilder_update_image_pipeline(imagePipelineArn, description,
  imageRecipeArn, infrastructureConfigurationArn,
  distributionConfigurationArn, imageTestsConfiguration, schedule, status,
  clientToken)
}
\arguments{
\item{imagePipelineArn}{[required] The Amazon Resource Name (ARN) of the image pipeline that you wish to
update.}

\item{description}{The description of the image pipeline.}

\item{imageRecipeArn}{The Amazon Resource Name (ARN) of the image recipe that will be used to
configure images updated by this image pipeline.}

\item{infrastructureConfigurationArn}{The Amazon Resource Name (ARN) of the infrastructure configuration that
will be used to build images updated by this image pipeline.}

\item{distributionConfigurationArn}{The Amazon Resource Name (ARN) of the distribution configuration that
will be used to configure and distribute images updated by this image
pipeline.}

\item{imageTestsConfiguration}{The image test configuration of the image pipeline.}

\item{schedule}{The schedule of the image pipeline.}

\item{status}{The status of the image pipeline.}

\item{clientToken}{[required] The idempotency token used to make this request idempotent.}
}
\description{
Updates a new image pipeline. Image pipelines enable you to automate the
creation and distribution of images.
}
\section{Request syntax}{
\preformatted{svc$update_image_pipeline(
  imagePipelineArn = "string",
  description = "string",
  imageRecipeArn = "string",
  infrastructureConfigurationArn = "string",
  distributionConfigurationArn = "string",
  imageTestsConfiguration = list(
    imageTestsEnabled = TRUE|FALSE,
    timeoutMinutes = 123
  ),
  schedule = list(
    scheduleExpression = "string",
    pipelineExecutionStartCondition = "EXPRESSION_MATCH_ONLY"|"EXPRESSION_MATCH_AND_DEPENDENCY_UPDATES_AVAILABLE"
  ),
  status = "DISABLED"|"ENABLED",
  clientToken = "string"
)
}
}

\keyword{internal}
