% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_finalize_device_claim}
\alias{iot1clickdevicesservice_finalize_device_claim}
\title{Given a device ID, finalizes the claim request for the associated device}
\usage{
iot1clickdevicesservice_finalize_device_claim(DeviceId, Tags)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}

\item{Tags}{A collection of key/value pairs defining the resource tags. For example,
\{ "tags": \{"key1": "value1", "key2": "value2"\} \}. For more information,
see \href{https://d1.awsstatic.com/whitepapers/aws-tagging-best-practices.pdf}{AWS Tagging Strategies}.}
}
\description{
Given a device ID, finalizes the claim request for the associated
device.

Claiming a device consists of initiating a claim, then publishing a
device event, and finalizing the claim. For a device of type button, a
device event can be published by simply clicking the device.
}
\section{Request syntax}{
\preformatted{svc$finalize_device_claim(
  DeviceId = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
