% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_invoke_device_method}
\alias{iot1clickdevicesservice_invoke_device_method}
\title{Given a device ID, issues a request to invoke a named device method
(with possible parameters)}
\usage{
iot1clickdevicesservice_invoke_device_method(DeviceId, DeviceMethod,
  DeviceMethodParameters)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}

\item{DeviceMethod}{The device method to invoke.}

\item{DeviceMethodParameters}{A JSON encoded string containing the device method request parameters.}
}
\description{
Given a device ID, issues a request to invoke a named device method
(with possible parameters). See the \"Example POST\" code snippet below.
}
\section{Request syntax}{
\preformatted{svc$invoke_device_method(
  DeviceId = "string",
  DeviceMethod = list(
    DeviceType = "string",
    MethodName = "string"
  ),
  DeviceMethodParameters = "string"
)
}
}

\keyword{internal}
