% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_list_devices}
\alias{iot1clickdevicesservice_list_devices}
\title{Lists the 1-Click compatible devices associated with your AWS account}
\usage{
iot1clickdevicesservice_list_devices(DeviceType, MaxResults, NextToken)
}
\arguments{
\item{DeviceType}{The type of the device, such as "button".}

\item{MaxResults}{The maximum number of results to return per request. If not set, a
default value of 100 is used.}

\item{NextToken}{The token to retrieve the next set of results.}
}
\description{
Lists the 1-Click compatible devices associated with your AWS account.
}
\section{Request syntax}{
\preformatted{iot1clickdevicesservice$list_devices(
  DeviceType = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
