% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_tag_resource}
\alias{iot1clickdevicesservice_tag_resource}
\title{Associate a set of tags with an AWS IoT 1-Click resource}
\usage{
iot1clickdevicesservice_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] Identifies the AWS IoT 1-Click resource to which tags should be added.
This value is an Amazon Resource Name (ARN).}

\item{Tags}{[required] The tags to be assigned to the AWS IoT 1-Click resource.}
}
\description{
Associate a set of tags with an AWS IoT 1-Click resource. You can then
activate these user-defined tags so that they appear on the Billing and
Cost Management console for cost allocation tracking.
}
\section{Request syntax}{
\preformatted{iot1clickdevicesservice$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
