% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_untag_resource}
\alias{iot1clickdevicesservice_untag_resource}
\title{Removes the association of tags from an AWS IoT 1-Click resource}
\usage{
iot1clickdevicesservice_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The AWS IoT 1-Click resource the tags will be removed from. This value
is an Amazon Resource Name (ARN).}

\item{TagKeys}{[required] A list of tag keys. Existing tags of the resource whose keys are members
of this list will be removed from the AWS IoT 1-Click resource.}
}
\description{
Removes the association of tags from an AWS IoT 1-Click resource.
}
\section{Request syntax}{
\preformatted{iot1clickdevicesservice$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
