% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_list_projects}
\alias{iot1clickprojects_list_projects}
\title{Lists the AWS IoT 1-Click project(s) associated with your AWS account
and region}
\usage{
iot1clickprojects_list_projects(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return per request. If not set, a
default value of 100 is used.}
}
\description{
Lists the AWS IoT 1-Click project(s) associated with your AWS account
and region.
}
\section{Request syntax}{
\preformatted{svc$list_projects(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
