% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_accept_certificate_transfer}
\alias{iot_accept_certificate_transfer}
\title{Accepts a pending certificate transfer}
\usage{
iot_accept_certificate_transfer(certificateId, setAsActive)
}
\arguments{
\item{certificateId}{[required] The ID of the certificate. (The last part of the certificate ARN
contains the certificate ID.)}

\item{setAsActive}{Specifies whether the certificate is active.}
}
\description{
Accepts a pending certificate transfer. The default state of the
certificate is INACTIVE.

To check for pending certificate transfers, call
\code{\link[=iot_list_certificates]{list_certificates}} to enumerate your
certificates.
}
\section{Request syntax}{
\preformatted{svc$accept_certificate_transfer(
  certificateId = "string",
  setAsActive = TRUE|FALSE
)
}
}

\keyword{internal}
