% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_attach_principal_policy}
\alias{iot_attach_principal_policy}
\title{Attaches the specified policy to the specified principal (certificate or
other credential)}
\usage{
iot_attach_principal_policy(policyName, principal)
}
\arguments{
\item{policyName}{[required] The policy name.}

\item{principal}{[required] The principal, which can be a certificate ARN (as returned from the
CreateCertificate operation) or an Amazon Cognito ID.}
}
\description{
Attaches the specified policy to the specified principal (certificate or
other credential).

\strong{Note:} This API is deprecated. Please use AttachPolicy instead.
}
\section{Request syntax}{
\preformatted{svc$attach_principal_policy(
  policyName = "string",
  principal = "string"
)
}
}

\keyword{internal}
