% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_attach_security_profile}
\alias{iot_attach_security_profile}
\title{Associates a Device Defender security profile with a thing group or this
account}
\usage{
iot_attach_security_profile(securityProfileName,
  securityProfileTargetArn)
}
\arguments{
\item{securityProfileName}{[required] The security profile that is attached.}

\item{securityProfileTargetArn}{[required] The ARN of the target (thing group) to which the security profile is
attached.}
}
\description{
Associates a Device Defender security profile with a thing group or this
account. Each thing group or account can have up to five security
profiles associated with it.
}
\section{Request syntax}{
\preformatted{svc$attach_security_profile(
  securityProfileName = "string",
  securityProfileTargetArn = "string"
)
}
}

\keyword{internal}
