% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_cancel_audit_task}
\alias{iot_cancel_audit_task}
\title{Cancels an audit that is in progress}
\usage{
iot_cancel_audit_task(taskId)
}
\arguments{
\item{taskId}{[required] The ID of the audit you want to cancel. You can only cancel an audit
that is "IN\\_PROGRESS".}
}
\description{
Cancels an audit that is in progress. The audit can be either scheduled
or on-demand. If the audit is not in progress, an
"InvalidRequestException" occurs.
}
\section{Request syntax}{
\preformatted{svc$cancel_audit_task(
  taskId = "string"
)
}
}

\keyword{internal}
