% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_domain_configuration}
\alias{iot_create_domain_configuration}
\title{Creates a domain configuration}
\usage{
iot_create_domain_configuration(domainConfigurationName, domainName,
  serverCertificateArns, validationCertificateArn, authorizerConfig,
  serviceType, tags)
}
\arguments{
\item{domainConfigurationName}{[required] The name of the domain configuration. This value must be unique to a
region.}

\item{domainName}{The name of the domain.}

\item{serverCertificateArns}{The ARNs of the certificates that AWS IoT passes to the device during
the TLS handshake. Currently you can specify only one certificate ARN.
This value is not required for AWS-managed domains.}

\item{validationCertificateArn}{The certificate used to validate the server certificate and prove domain
name ownership. This certificate must be signed by a public certificate
authority. This value is not required for AWS-managed domains.}

\item{authorizerConfig}{An object that specifies the authorization service for a domain.}

\item{serviceType}{The type of service delivered by the endpoint.

AWS IoT Core currently supports only the \code{DATA} service type.}

\item{tags}{Metadata which can be used to manage the domain configuration.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags
"key1=value1&key2=value2..."

For the cli-input-json file use format: "tags":
"key1=value1&key2=value2..."}
}
\description{
Creates a domain configuration.
}
\details{
The domain configuration feature is in public preview and is subject to
change.
}
\section{Request syntax}{
\preformatted{svc$create_domain_configuration(
  domainConfigurationName = "string",
  domainName = "string",
  serverCertificateArns = list(
    "string"
  ),
  validationCertificateArn = "string",
  authorizerConfig = list(
    defaultAuthorizerName = "string",
    allowAuthorizerOverride = TRUE|FALSE
  ),
  serviceType = "DATA"|"CREDENTIAL_PROVIDER"|"JOBS",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
