% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_keys_and_certificate}
\alias{iot_create_keys_and_certificate}
\title{Creates a 2048-bit RSA key pair and issues an X}
\usage{
iot_create_keys_and_certificate(setAsActive)
}
\arguments{
\item{setAsActive}{Specifies whether the certificate is active.}
}
\description{
Creates a 2048-bit RSA key pair and issues an X.509 certificate using
the issued public key.
}
\details{
\strong{Note} This is the only time AWS IoT issues the private key for this
certificate, so it is important to keep it in a secure location.
}
\section{Request syntax}{
\preformatted{iot$create_keys_and_certificate(
  setAsActive = TRUE|FALSE
)
}
}

\keyword{internal}
