% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_mitigation_action}
\alias{iot_create_mitigation_action}
\title{Defines an action that can be applied to audit findings by using
StartAuditMitigationActionsTask}
\usage{
iot_create_mitigation_action(actionName, roleArn, actionParams, tags)
}
\arguments{
\item{actionName}{[required] A friendly name for the action. Choose a friendly name that accurately
describes the action (for example, \code{EnableLoggingAction}).}

\item{roleArn}{[required] The ARN of the IAM role that is used to apply the mitigation action.}

\item{actionParams}{[required] Defines the type of action and the parameters for that action.}

\item{tags}{Metadata that can be used to manage the mitigation action.}
}
\description{
Defines an action that can be applied to audit findings by using
StartAuditMitigationActionsTask. Only certain types of mitigation
actions can be applied to specific check names. For more information,
see \href{https://docs.aws.amazon.com/iot/latest/developerguide/}{Mitigation actions}. Each
mitigation action can apply only one type of change.
}
\section{Request syntax}{
\preformatted{svc$create_mitigation_action(
  actionName = "string",
  roleArn = "string",
  actionParams = list(
    updateDeviceCertificateParams = list(
      action = "DEACTIVATE"
    ),
    updateCACertificateParams = list(
      action = "DEACTIVATE"
    ),
    addThingsToThingGroupParams = list(
      thingGroupNames = list(
        "string"
      ),
      overrideDynamicGroups = TRUE|FALSE
    ),
    replaceDefaultPolicyVersionParams = list(
      templateName = "BLANK_POLICY"
    ),
    enableIoTLoggingParams = list(
      roleArnForLogging = "string",
      logLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED"
    ),
    publishFindingToSnsParams = list(
      topicArn = "string"
    )
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
