% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_policy}
\alias{iot_create_policy}
\title{Creates an AWS IoT policy}
\usage{
iot_create_policy(policyName, policyDocument, tags)
}
\arguments{
\item{policyName}{[required] The policy name.}

\item{policyDocument}{[required] The JSON document that describes the policy. \strong{policyDocument} must
have a minimum length of 1, with a maximum length of 2048, excluding
whitespace.}

\item{tags}{Metadata which can be used to manage the policy.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags
"key1=value1&key2=value2..."

For the cli-input-json file use format: "tags":
"key1=value1&key2=value2..."}
}
\description{
Creates an AWS IoT policy.
}
\details{
The created policy is the default version for the policy. This operation
creates a policy version with a version identifier of \strong{1} and sets
\strong{1} as the policy's default version.
}
\section{Request syntax}{
\preformatted{svc$create_policy(
  policyName = "string",
  policyDocument = "string",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
