% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_scheduled_audit}
\alias{iot_create_scheduled_audit}
\title{Creates a scheduled audit that is run at a specified time interval}
\usage{
iot_create_scheduled_audit(frequency, dayOfMonth, dayOfWeek,
  targetCheckNames, scheduledAuditName, tags)
}
\arguments{
\item{frequency}{[required] How often the scheduled audit takes place, either \code{DAILY}, \code{WEEKLY},
\code{BIWEEKLY} or \code{MONTHLY}. The start time of each audit is determined by
the system.}

\item{dayOfMonth}{The day of the month on which the scheduled audit takes place. This can
be "1" through "31" or "LAST". This field is required if the "frequency"
parameter is set to \code{MONTHLY}. If days 29 to 31 are specified, and the
month doesn't have that many days, the audit takes place on the \code{LAST}
day of the month.}

\item{dayOfWeek}{The day of the week on which the scheduled audit takes place, either
\code{SUN}, \code{MON}, \code{TUE}, \code{WED}, \code{THU}, \code{FRI}, or \code{SAT}. This field is
required if the \code{frequency} parameter is set to \code{WEEKLY} or \code{BIWEEKLY}.}

\item{targetCheckNames}{[required] Which checks are performed during the scheduled audit. Checks must be
enabled for your account. (Use
\code{\link[=iot_describe_account_audit_configuration]{describe_account_audit_configuration}}
to see the list of all checks, including those that are enabled or use
\code{\link[=iot_update_account_audit_configuration]{update_account_audit_configuration}}
to select which checks are enabled.)}

\item{scheduledAuditName}{[required] The name you want to give to the scheduled audit. (Max. 128 chars)}

\item{tags}{Metadata that can be used to manage the scheduled audit.}
}
\description{
Creates a scheduled audit that is run at a specified time interval.
}
\section{Request syntax}{
\preformatted{svc$create_scheduled_audit(
  frequency = "DAILY"|"WEEKLY"|"BIWEEKLY"|"MONTHLY",
  dayOfMonth = "string",
  dayOfWeek = "SUN"|"MON"|"TUE"|"WED"|"THU"|"FRI"|"SAT",
  targetCheckNames = list(
    "string"
  ),
  scheduledAuditName = "string",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
