% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_security_profile}
\alias{iot_create_security_profile}
\title{Creates a Device Defender security profile}
\usage{
iot_create_security_profile(securityProfileName,
  securityProfileDescription, behaviors, alertTargets,
  additionalMetricsToRetain, additionalMetricsToRetainV2, tags)
}
\arguments{
\item{securityProfileName}{[required] The name you are giving to the security profile.}

\item{securityProfileDescription}{A description of the security profile.}

\item{behaviors}{Specifies the behaviors that, when violated by a device (thing), cause
an alert.}

\item{alertTargets}{Specifies the destinations to which alerts are sent. (Alerts are always
sent to the console.) Alerts are generated when a device (thing)
violates a behavior.}

\item{additionalMetricsToRetain}{A list of metrics whose data is retained (stored). By default, data is
retained for any metric used in the profile's \code{behaviors}, but it is
also retained for any metric specified here.

\strong{Note:} This API field is deprecated. Please use
CreateSecurityProfileRequest$additionalMetricsToRetainV2 instead.}

\item{additionalMetricsToRetainV2}{A list of metrics whose data is retained (stored). By default, data is
retained for any metric used in the profile's \code{behaviors}, but it is
also retained for any metric specified here.}

\item{tags}{Metadata that can be used to manage the security profile.}
}
\description{
Creates a Device Defender security profile.
}
\section{Request syntax}{
\preformatted{svc$create_security_profile(
  securityProfileName = "string",
  securityProfileDescription = "string",
  behaviors = list(
    list(
      name = "string",
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      ),
      criteria = list(
        comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set",
        value = list(
          count = 123,
          cidrs = list(
            "string"
          ),
          ports = list(
            123
          )
        ),
        durationSeconds = 123,
        consecutiveDatapointsToAlarm = 123,
        consecutiveDatapointsToClear = 123,
        statisticalThreshold = list(
          statistic = "string"
        )
      )
    )
  ),
  alertTargets = list(
    list(
      alertTargetArn = "string",
      roleArn = "string"
    )
  ),
  additionalMetricsToRetain = list(
    "string"
  ),
  additionalMetricsToRetainV2 = list(
    list(
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      )
    )
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
