% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_thing}
\alias{iot_create_thing}
\title{Creates a thing record in the registry}
\usage{
iot_create_thing(thingName, thingTypeName, attributePayload,
  billingGroupName)
}
\arguments{
\item{thingName}{[required] The name of the thing to create.

You can\'t change a thing\'s name after you create it. To change a
thing\'s name, you must create a new thing, give it the new name, and
then delete the old thing.}

\item{thingTypeName}{The name of the thing type associated with the new thing.}

\item{attributePayload}{The attribute payload, which consists of up to three name/value pairs in
a JSON document. For example:

\verb{\\\{\\"attributes\\":\\\{\\"string1\\":\\"string2\\"\\\}\\\}}}

\item{billingGroupName}{The name of the billing group the thing will be added to.}
}
\description{
Creates a thing record in the registry. If this call is made multiple
times using the same thing name and configuration, the call will
succeed. If this call is made with the same thing name but different
configuration a \code{ResourceAlreadyExistsException} is thrown.
}
\details{
This is a control plane operation. See
\href{https://docs.aws.amazon.com/iot/latest/developerguide/iot-authorization.html}{Authorization}
for information about authorizing control plane actions.
}
\section{Request syntax}{
\preformatted{svc$create_thing(
  thingName = "string",
  thingTypeName = "string",
  attributePayload = list(
    attributes = list(
      "string"
    ),
    merge = TRUE|FALSE
  ),
  billingGroupName = "string"
)
}
}

\keyword{internal}
