% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_thing_group}
\alias{iot_create_thing_group}
\title{Create a thing group}
\usage{
iot_create_thing_group(thingGroupName, parentGroupName,
  thingGroupProperties, tags)
}
\arguments{
\item{thingGroupName}{[required] The thing group name to create.}

\item{parentGroupName}{The name of the parent thing group.}

\item{thingGroupProperties}{The thing group properties.}

\item{tags}{Metadata which can be used to manage the thing group.}
}
\description{
Create a thing group.
}
\details{
This is a control plane operation. See
\href{http://docs.aws.amazon.com/iot/latest/developerguide/authorization.html}{Authorization}
for information about authorizing control plane actions.
}
\section{Request syntax}{
\preformatted{iot$create_thing_group(
  thingGroupName = "string",
  parentGroupName = "string",
  thingGroupProperties = list(
    thingGroupDescription = "string",
    attributePayload = list(
      attributes = list(
        "string"
      ),
      merge = TRUE|FALSE
    )
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
