% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_authorizer}
\alias{iot_delete_authorizer}
\title{Deletes an authorizer}
\usage{
iot_delete_authorizer(authorizerName)
}
\arguments{
\item{authorizerName}{[required] The name of the authorizer to delete.}
}
\description{
Deletes an authorizer.
}
\section{Request syntax}{
\preformatted{iot$delete_authorizer(
  authorizerName = "string"
)
}
}

\keyword{internal}
