% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_role_alias}
\alias{iot_delete_role_alias}
\title{Deletes a role alias}
\usage{
iot_delete_role_alias(roleAlias)
}
\arguments{
\item{roleAlias}{[required] The role alias to delete.}
}
\description{
Deletes a role alias
}
\section{Request syntax}{
\preformatted{iot$delete_role_alias(
  roleAlias = "string"
)
}
}

\keyword{internal}
