% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_security_profile}
\alias{iot_delete_security_profile}
\title{Deletes a Device Defender security profile}
\usage{
iot_delete_security_profile(securityProfileName, expectedVersion)
}
\arguments{
\item{securityProfileName}{[required] The name of the security profile to be deleted.}

\item{expectedVersion}{The expected version of the security profile. A new version is generated
whenever the security profile is updated. If you specify a value that is
different than the actual version, a \code{VersionConflictException} is
thrown.}
}
\description{
Deletes a Device Defender security profile.
}
\section{Request syntax}{
\preformatted{iot$delete_security_profile(
  securityProfileName = "string",
  expectedVersion = 123
)
}
}

\keyword{internal}
