% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_account_audit_configuration}
\alias{iot_describe_account_audit_configuration}
\title{Gets information about the Device Defender audit settings for this
account}
\usage{
iot_describe_account_audit_configuration()
}
\description{
Gets information about the Device Defender audit settings for this
account. Settings include how audit notifications are sent and which
audit checks are enabled or disabled.
}
\section{Request syntax}{
\preformatted{iot$describe_account_audit_configuration()
}
}

\keyword{internal}
