% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_audit_finding}
\alias{iot_describe_audit_finding}
\title{Gets information about a single audit finding}
\usage{
iot_describe_audit_finding(findingId)
}
\arguments{
\item{findingId}{[required] A unique identifier for a single audit finding. You can use this
identifier to apply mitigation actions to the finding.}
}
\description{
Gets information about a single audit finding. Properties include the
reason for noncompliance, the severity of the issue, and the start time
when the audit that returned the finding.
}
\section{Request syntax}{
\preformatted{svc$describe_audit_finding(
  findingId = "string"
)
}
}

\keyword{internal}
