% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_audit_mitigation_actions_task}
\alias{iot_describe_audit_mitigation_actions_task}
\title{Gets information about an audit mitigation task that is used to apply
mitigation actions to a set of audit findings}
\usage{
iot_describe_audit_mitigation_actions_task(taskId)
}
\arguments{
\item{taskId}{[required] The unique identifier for the audit mitigation task.}
}
\description{
Gets information about an audit mitigation task that is used to apply
mitigation actions to a set of audit findings. Properties include the
actions being applied, the audit checks to which they're being applied,
the task status, and aggregated task statistics.
}
\section{Request syntax}{
\preformatted{svc$describe_audit_mitigation_actions_task(
  taskId = "string"
)
}
}

\keyword{internal}
