% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_endpoint}
\alias{iot_describe_endpoint}
\title{Returns a unique endpoint specific to the AWS account making the call}
\usage{
iot_describe_endpoint(endpointType)
}
\arguments{
\item{endpointType}{The endpoint type. Valid endpoint types include:
\itemize{
\item \code{iot:Data} - Returns a VeriSign signed data endpoint.
}<!-- -->

\itemize{
\item \code{iot:Data-ATS} - Returns an ATS signed data endpoint.
}<!-- -->

\itemize{
\item \code{iot:CredentialProvider} - Returns an AWS IoT credentials provider
API endpoint.
}<!-- -->

\itemize{
\item \code{iot:Jobs} - Returns an AWS IoT device management Jobs API endpoint.
}}
}
\description{
Returns a unique endpoint specific to the AWS account making the call.
}
\section{Request syntax}{
\preformatted{svc$describe_endpoint(
  endpointType = "string"
)
}
}

\keyword{internal}
