% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_job}
\alias{iot_describe_job}
\title{Describes a job}
\usage{
iot_describe_job(jobId)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}
}
\description{
Describes a job.
}
\section{Request syntax}{
\preformatted{svc$describe_job(
  jobId = "string"
)
}
}

\keyword{internal}
