% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_mitigation_action}
\alias{iot_describe_mitigation_action}
\title{Gets information about a mitigation action}
\usage{
iot_describe_mitigation_action(actionName)
}
\arguments{
\item{actionName}{[required] The friendly name that uniquely identifies the mitigation action.}
}
\description{
Gets information about a mitigation action.
}
\section{Request syntax}{
\preformatted{svc$describe_mitigation_action(
  actionName = "string"
)
}
}

\keyword{internal}
