% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_thing}
\alias{iot_describe_thing}
\title{Gets information about the specified thing}
\usage{
iot_describe_thing(thingName)
}
\arguments{
\item{thingName}{[required] The name of the thing.}
}
\description{
Gets information about the specified thing.
}
\section{Request syntax}{
\preformatted{iot$describe_thing(
  thingName = "string"
)
}
}

\keyword{internal}
