% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_policy}
\alias{iot_get_policy}
\title{Gets information about the specified policy with the policy document of
the default version}
\usage{
iot_get_policy(policyName)
}
\arguments{
\item{policyName}{[required] The name of the policy.}
}
\description{
Gets information about the specified policy with the policy document of
the default version.
}
\section{Request syntax}{
\preformatted{svc$get_policy(
  policyName = "string"
)
}
}

\keyword{internal}
