% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_topic_rule_destination}
\alias{iot_get_topic_rule_destination}
\title{Gets information about a topic rule destination}
\usage{
iot_get_topic_rule_destination(arn)
}
\arguments{
\item{arn}{[required] The ARN of the topic rule destination.}
}
\description{
Gets information about a topic rule destination.
}
\section{Request syntax}{
\preformatted{svc$get_topic_rule_destination(
  arn = "string"
)
}
}

\keyword{internal}
