% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_v2_logging_options}
\alias{iot_get_v2_logging_options}
\title{Gets the fine grained logging options}
\usage{
iot_get_v2_logging_options()
}
\description{
Gets the fine grained logging options.
}
\section{Request syntax}{
\preformatted{iot$get_v2_logging_options()
}
}

\keyword{internal}
