% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_audit_tasks}
\alias{iot_list_audit_tasks}
\title{Lists the Device Defender audits that have been performed during a given
time period}
\usage{
iot_list_audit_tasks(startTime, endTime, taskType, taskStatus,
  nextToken, maxResults)
}
\arguments{
\item{startTime}{[required] The beginning of the time period. Audit information is retained for a
limited time (90 days). Requesting a start time prior to what is
retained results in an "InvalidRequestException".}

\item{endTime}{[required] The end of the time period.}

\item{taskType}{A filter to limit the output to the specified type of audit: can be one
of "ON\\_DEMAND\\_AUDIT\\\emph{TASK" or "SCHEDULED\\}\\_AUDIT\\_TASK".}

\item{taskStatus}{A filter to limit the output to audits with the specified completion
status: can be one of "IN\\_PROGRESS", "COMPLETED", "FAILED", or
"CANCELED".}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}
}
\description{
Lists the Device Defender audits that have been performed during a given
time period.
}
\section{Request syntax}{
\preformatted{svc$list_audit_tasks(
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  taskType = "ON_DEMAND_AUDIT_TASK"|"SCHEDULED_AUDIT_TASK",
  taskStatus = "IN_PROGRESS"|"COMPLETED"|"FAILED"|"CANCELED",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
