% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_domain_configurations}
\alias{iot_list_domain_configurations}
\title{Gets a list of domain configurations for the user}
\usage{
iot_list_domain_configurations(marker, pageSize, serviceType)
}
\arguments{
\item{marker}{The marker for the next set of results.}

\item{pageSize}{The result page size.}

\item{serviceType}{The type of service delivered by the endpoint.}
}
\description{
Gets a list of domain configurations for the user. This list is sorted
alphabetically by domain configuration name.

The domain configuration feature is in public preview and is subject to
change.
}
\section{Request syntax}{
\preformatted{svc$list_domain_configurations(
  marker = "string",
  pageSize = 123,
  serviceType = "DATA"|"CREDENTIAL_PROVIDER"|"JOBS"
)
}
}

\keyword{internal}
