% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_indices}
\alias{iot_list_indices}
\title{Lists the search indices}
\usage{
iot_list_indices(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token used to get the next set of results, or \code{null} if there are no
additional results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the search indices.
}
\section{Request syntax}{
\preformatted{svc$list_indices(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
