% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_principal_things}
\alias{iot_list_principal_things}
\title{Lists the things associated with the specified principal}
\usage{
iot_list_principal_things(nextToken, maxResults, principal)
}
\arguments{
\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return in this operation.}

\item{principal}{[required] The principal.}
}
\description{
Lists the things associated with the specified principal.
}
\section{Request syntax}{
\preformatted{iot$list_principal_things(
  nextToken = "string",
  maxResults = 123,
  principal = "string"
)
}
}

\keyword{internal}
