% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_streams}
\alias{iot_list_streams}
\title{Lists all of the streams in your AWS account}
\usage{
iot_list_streams(maxResults, nextToken, ascendingOrder)
}
\arguments{
\item{maxResults}{The maximum number of results to return at a time.}

\item{nextToken}{A token used to get the next set of results.}

\item{ascendingOrder}{Set to true to return the list of streams in ascending order.}
}
\description{
Lists all of the streams in your AWS account.
}
\section{Request syntax}{
\preformatted{iot$list_streams(
  maxResults = 123,
  nextToken = "string",
  ascendingOrder = TRUE|FALSE
)
}
}

\keyword{internal}
