% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_targets_for_security_profile}
\alias{iot_list_targets_for_security_profile}
\title{Lists the targets (thing groups) associated with a given Device Defender
security profile}
\usage{
iot_list_targets_for_security_profile(securityProfileName, nextToken,
  maxResults)
}
\arguments{
\item{securityProfileName}{[required] The security profile.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the targets (thing groups) associated with a given Device Defender
security profile.
}
\section{Request syntax}{
\preformatted{iot$list_targets_for_security_profile(
  securityProfileName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
