% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_register_thing}
\alias{iot_register_thing}
\title{Provisions a thing}
\usage{
iot_register_thing(templateBody, parameters)
}
\arguments{
\item{templateBody}{[required] The provisioning template. See \href{http://docs.aws.amazon.com/iot/latest/developerguide/programmatic-provisioning.html}{Programmatic Provisioning}
for more information.}

\item{parameters}{The parameters for provisioning a thing. See \href{http://docs.aws.amazon.com/iot/latest/developerguide/programmatic-provisioning.html}{Programmatic Provisioning}
for more information.}
}
\description{
Provisions a thing.
}
\section{Request syntax}{
\preformatted{iot$register_thing(
  templateBody = "string",
  parameters = list(
    "string"
  )
)
}
}

\keyword{internal}
