% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_transfer_certificate}
\alias{iot_transfer_certificate}
\title{Transfers the specified certificate to the specified AWS account}
\usage{
iot_transfer_certificate(certificateId, targetAwsAccount,
  transferMessage)
}
\arguments{
\item{certificateId}{[required] The ID of the certificate. (The last part of the certificate ARN
contains the certificate ID.)}

\item{targetAwsAccount}{[required] The AWS account.}

\item{transferMessage}{The transfer message.}
}
\description{
Transfers the specified certificate to the specified AWS account.
}
\details{
You can cancel the transfer until it is acknowledged by the recipient.

No notification is sent to the transfer destination's account. It is up
to the caller to notify the transfer target.

The certificate being transferred must not be in the ACTIVE state. You
can use the UpdateCertificate API to deactivate it.

The certificate must not have any policies attached to it. You can use
the DetachPrincipalPolicy API to detach them.
}
\section{Request syntax}{
\preformatted{iot$transfer_certificate(
  certificateId = "string",
  targetAwsAccount = "string",
  transferMessage = "string"
)
}
}

\keyword{internal}
