% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_untag_resource}
\alias{iot_untag_resource}
\title{Removes the given tags (metadata) from the resource}
\usage{
iot_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tagKeys}{[required] A list of the keys of the tags to be removed from the resource.}
}
\description{
Removes the given tags (metadata) from the resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
