% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_account_audit_configuration}
\alias{iot_update_account_audit_configuration}
\title{Configures or reconfigures the Device Defender audit settings for this
account}
\usage{
iot_update_account_audit_configuration(roleArn,
  auditNotificationTargetConfigurations, auditCheckConfigurations)
}
\arguments{
\item{roleArn}{The Amazon Resource Name (ARN) of the role that grants permission to AWS
IoT to access information about your devices, policies, certificates,
and other items as required when performing an audit.}

\item{auditNotificationTargetConfigurations}{Information about the targets to which audit notifications are sent.}

\item{auditCheckConfigurations}{Specifies which audit checks are enabled and disabled for this account.
Use
\code{\link[=iot_describe_account_audit_configuration]{describe_account_audit_configuration}}
to see the list of all checks, including those that are currently
enabled.

Some data collection might start immediately when certain checks are
enabled. When a check is disabled, any data collected so far in relation
to the check is deleted.

You cannot disable a check if it's used by any scheduled audit. You must
first delete the check from the scheduled audit or delete the scheduled
audit itself.

On the first call to
\code{\link[=iot_update_account_audit_configuration]{update_account_audit_configuration}},
this parameter is required and must specify at least one enabled check.}
}
\description{
Configures or reconfigures the Device Defender audit settings for this
account. Settings include how audit notifications are sent and which
audit checks are enabled or disabled.
}
\section{Request syntax}{
\preformatted{svc$update_account_audit_configuration(
  roleArn = "string",
  auditNotificationTargetConfigurations = list(
    list(
      targetArn = "string",
      roleArn = "string",
      enabled = TRUE|FALSE
    )
  ),
  auditCheckConfigurations = list(
    list(
      enabled = TRUE|FALSE
    )
  )
)
}
}

\keyword{internal}
