% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_authorizer}
\alias{iot_update_authorizer}
\title{Updates an authorizer}
\usage{
iot_update_authorizer(authorizerName, authorizerFunctionArn,
  tokenKeyName, tokenSigningPublicKeys, status)
}
\arguments{
\item{authorizerName}{[required] The authorizer name.}

\item{authorizerFunctionArn}{The ARN of the authorizer\'s Lambda function.}

\item{tokenKeyName}{The key used to extract the token from the HTTP headers.}

\item{tokenSigningPublicKeys}{The public keys used to verify the token signature.}

\item{status}{The status of the update authorizer request.}
}
\description{
Updates an authorizer.
}
\section{Request syntax}{
\preformatted{svc$update_authorizer(
  authorizerName = "string",
  authorizerFunctionArn = "string",
  tokenKeyName = "string",
  tokenSigningPublicKeys = list(
    "string"
  ),
  status = "ACTIVE"|"INACTIVE"
)
}
}

\keyword{internal}
