% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_dimension}
\alias{iot_update_dimension}
\title{Updates the definition for a dimension}
\usage{
iot_update_dimension(name, stringValues)
}
\arguments{
\item{name}{[required] A unique identifier for the dimension. Choose something that describes
the type and value to make it easy to remember what it does.}

\item{stringValues}{[required] Specifies the value or list of values for the dimension. For
\code{TOPIC_FILTER} dimensions, this is a pattern used to match the MQTT
topic (for example, "admin/\\#").}
}
\description{
Updates the definition for a dimension. You cannot change the type of a
dimension after it is created (you can delete it and recreate it).
}
\section{Request syntax}{
\preformatted{svc$update_dimension(
  name = "string",
  stringValues = list(
    "string"
  )
)
}
}

\keyword{internal}
