% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_security_profile}
\alias{iot_update_security_profile}
\title{Updates a Device Defender security profile}
\usage{
iot_update_security_profile(securityProfileName,
  securityProfileDescription, behaviors, alertTargets, expectedVersion)
}
\arguments{
\item{securityProfileName}{[required] The name of the security profile you want to update.}

\item{securityProfileDescription}{A description of the security profile.}

\item{behaviors}{Specifies the behaviors that, when violated by a device (thing), cause
an alert.}

\item{alertTargets}{Where the alerts are sent. (Alerts are always sent to the console.)}

\item{expectedVersion}{The expected version of the security profile. A new version is generated
whenever the security profile is updated. If you specify a value that is
different than the actual version, a \code{VersionConflictException} is
thrown.}
}
\description{
Updates a Device Defender security profile.
}
\section{Request syntax}{
\preformatted{iot$update_security_profile(
  securityProfileName = "string",
  securityProfileDescription = "string",
  behaviors = list(
    list(
      name = "string",
      metric = "string",
      criteria = list(
        comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set",
        value = list(
          count = 123,
          cidrs = list(
            "string"
          ),
          ports = list(
            123
          )
        ),
        durationSeconds = 123
      )
    )
  ),
  alertTargets = list(
    list(
      alertTargetArn = "string",
      roleArn = "string"
    )
  ),
  expectedVersion = 123
)
}
}

\keyword{internal}
