% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_thing_groups_for_thing}
\alias{iot_update_thing_groups_for_thing}
\title{Updates the groups to which the thing belongs}
\usage{
iot_update_thing_groups_for_thing(thingName, thingGroupsToAdd,
  thingGroupsToRemove, overrideDynamicGroups)
}
\arguments{
\item{thingName}{The thing whose group memberships will be updated.}

\item{thingGroupsToAdd}{The groups to which the thing will be added.}

\item{thingGroupsToRemove}{The groups from which the thing will be removed.}

\item{overrideDynamicGroups}{Override dynamic thing groups with static thing groups when 10-group
limit is reached. If a thing belongs to 10 thing groups, and one or more
of those groups are dynamic thing groups, adding a thing to a static
group removes the thing from the last dynamic group.}
}
\description{
Updates the groups to which the thing belongs.
}
\section{Request syntax}{
\preformatted{iot$update_thing_groups_for_thing(
  thingName = "string",
  thingGroupsToAdd = list(
    "string"
  ),
  thingGroupsToRemove = list(
    "string"
  ),
  overrideDynamicGroups = TRUE|FALSE
)
}
}

\keyword{internal}
