% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_describe_dataset}
\alias{iotanalytics_describe_dataset}
\title{Retrieves information about a data set}
\usage{
iotanalytics_describe_dataset(datasetName)
}
\arguments{
\item{datasetName}{[required] The name of the data set whose information is retrieved.}
}
\description{
Retrieves information about a data set.
}
\section{Request syntax}{
\preformatted{iotanalytics$describe_dataset(
  datasetName = "string"
)
}
}

\keyword{internal}
