% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_get_dataset_content}
\alias{iotanalytics_get_dataset_content}
\title{Retrieves the contents of a data set as pre-signed URIs}
\usage{
iotanalytics_get_dataset_content(datasetName, versionId)
}
\arguments{
\item{datasetName}{[required] The name of the data set whose contents are retrieved.}

\item{versionId}{The version of the data set whose contents are retrieved. You can also
use the strings "\$LATEST" or "\$LATEST\_SUCCEEDED" to retrieve the
contents of the latest or latest successfully completed data set. If not
specified, "\$LATEST\_SUCCEEDED" is the default.}
}
\description{
Retrieves the contents of a data set as pre-signed URIs.
}
\section{Request syntax}{
\preformatted{iotanalytics$get_dataset_content(
  datasetName = "string",
  versionId = "string"
)
}
}

\keyword{internal}
