% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_list_dataset_contents}
\alias{iotanalytics_list_dataset_contents}
\title{Lists information about data set contents that have been created}
\usage{
iotanalytics_list_dataset_contents(datasetName, nextToken, maxResults,
  scheduledOnOrAfter, scheduledBefore)
}
\arguments{
\item{datasetName}{[required] The name of the data set whose contents information you want to list.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in this request.}

\item{scheduledOnOrAfter}{A filter to limit results to those data set contents whose creation is
scheduled on or after the given time. See the field \code{triggers.schedule}
in the \code{\link[=iotanalytics_create_dataset]{create_dataset}} request.
(timestamp)}

\item{scheduledBefore}{A filter to limit results to those data set contents whose creation is
scheduled before the given time. See the field \code{triggers.schedule} in
the \code{\link[=iotanalytics_create_dataset]{create_dataset}} request. (timestamp)}
}
\description{
Lists information about data set contents that have been created.
}
\section{Request syntax}{
\preformatted{svc$list_dataset_contents(
  datasetName = "string",
  nextToken = "string",
  maxResults = 123,
  scheduledOnOrAfter = as.POSIXct(
    "2015-01-01"
  ),
  scheduledBefore = as.POSIXct(
    "2015-01-01"
  )
)
}
}

\keyword{internal}
