% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_list_datasets}
\alias{iotanalytics_list_datasets}
\title{Retrieves information about data sets}
\usage{
iotanalytics_list_datasets(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in this request.

The default value is 100.}
}
\description{
Retrieves information about data sets.
}
\section{Request syntax}{
\preformatted{iotanalytics$list_datasets(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
