% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_update_channel}
\alias{iotanalytics_update_channel}
\title{Updates the settings of a channel}
\usage{
iotanalytics_update_channel(channelName, retentionPeriod)
}
\arguments{
\item{channelName}{[required] The name of the channel to be updated.}

\item{retentionPeriod}{How long, in days, message data is kept for the channel.}
}
\description{
Updates the settings of a channel.
}
\section{Request syntax}{
\preformatted{iotanalytics$update_channel(
  channelName = "string",
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  )
)
}
}

\keyword{internal}
