% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_update_dataset}
\alias{iotanalytics_update_dataset}
\title{Updates the settings of a data set}
\usage{
iotanalytics_update_dataset(datasetName, actions, triggers,
  contentDeliveryRules, retentionPeriod, versioningConfiguration)
}
\arguments{
\item{datasetName}{[required] The name of the data set to update.}

\item{actions}{[required] A list of "DatasetAction" objects.}

\item{triggers}{A list of "DatasetTrigger" objects. The list can be empty or can
contain up to five \strong{DataSetTrigger} objects.}

\item{contentDeliveryRules}{When data set contents are created they are delivered to destinations
specified here.}

\item{retentionPeriod}{How long, in days, data set contents are kept for the data set.}

\item{versioningConfiguration}{[Optional] How many versions of data set contents are kept. If not
specified or set to null, only the latest version plus the latest
succeeded version (if they are different) are kept for the time period
specified by the "retentionPeriod" parameter. (For more information,
see
https://docs.aws.amazon.com/iotanalytics/latest/userguide/getting-started.html\#aws-iot-analytics-dataset-versions)}
}
\description{
Updates the settings of a data set.
}
\section{Request syntax}{
\preformatted{iotanalytics$update_dataset(
  datasetName = "string",
  actions = list(
    list(
      actionName = "string",
      queryAction = list(
        sqlQuery = "string",
        filters = list(
          list(
            deltaTime = list(
              offsetSeconds = 123,
              timeExpression = "string"
            )
          )
        )
      ),
      containerAction = list(
        image = "string",
        executionRoleArn = "string",
        resourceConfiguration = list(
          computeType = "ACU_1"|"ACU_2",
          volumeSizeInGB = 123
        ),
        variables = list(
          list(
            name = "string",
            stringValue = "string",
            doubleValue = 123.0,
            datasetContentVersionValue = list(
              datasetName = "string"
            ),
            outputFileUriValue = list(
              fileName = "string"
            )
          )
        )
      )
    )
  ),
  triggers = list(
    list(
      schedule = list(
        expression = "string"
      ),
      dataset = list(
        name = "string"
      )
    )
  ),
  contentDeliveryRules = list(
    list(
      entryName = "string",
      destination = list(
        iotEventsDestinationConfiguration = list(
          inputName = "string",
          roleArn = "string"
        )
      )
    )
  ),
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  ),
  versioningConfiguration = list(
    unlimited = TRUE|FALSE,
    maxVersions = 123
  )
)
}
}

\keyword{internal}
