% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_update_datastore}
\alias{iotanalytics_update_datastore}
\title{Updates the settings of a data store}
\usage{
iotanalytics_update_datastore(datastoreName, retentionPeriod,
  datastoreStorage)
}
\arguments{
\item{datastoreName}{[required] The name of the data store to be updated.}

\item{retentionPeriod}{How long, in days, message data is kept for the data store. The
retention period cannot be updated if the data store's S3 storage is
customer-managed.}

\item{datastoreStorage}{Where data store data is stored. You may choose one of
"serviceManagedS3" or "customerManagedS3" storage. If not specified, the
default is "serviceManagedS3". This cannot be changed after the data
store is created.}
}
\description{
Updates the settings of a data store.
}
\section{Request syntax}{
\preformatted{svc$update_datastore(
  datastoreName = "string",
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  ),
  datastoreStorage = list(
    serviceManagedS3 = list(),
    customerManagedS3 = list(
      bucket = "string",
      keyPrefix = "string",
      roleArn = "string"
    )
  )
)
}
}

\keyword{internal}
