% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{iotdataplane}
\alias{iotdataplane}
\title{AWS IoT Data Plane}
\usage{
iotdataplane()
}
\description{
AWS IoT

AWS IoT-Data enables secure, bi-directional communication between
Internet-connected things (such as sensors, actuators, embedded devices,
or smart appliances) and the AWS cloud. It implements a broker for
applications and things to publish messages over HTTP (Publish) and
retrieve, update, and delete thing shadows. A thing shadow is a
persistent representation of your things and their state in the AWS
cloud.
}
\section{Operations}{

\tabular{ll}{
\link[=iotdataplane_delete_thing_shadow]{delete_thing_shadow} \tab Deletes the thing shadow for the specified thing\cr
\link[=iotdataplane_get_thing_shadow]{get_thing_shadow} \tab Gets the thing shadow for the specified thing \cr
\link[=iotdataplane_publish]{publish} \tab Publishes state information \cr
\link[=iotdataplane_update_thing_shadow]{update_thing_shadow} \tab Updates the thing shadow for the specified thing
}
}

\examples{
\donttest{svc <- iotdataplane()
svc$delete_thing_shadow(
  Foo = 123
)}

}
