% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdataplane_operations.R
\name{iotdataplane_update_thing_shadow}
\alias{iotdataplane_update_thing_shadow}
\title{Updates the shadow for the specified thing}
\usage{
iotdataplane_update_thing_shadow(thingName, shadowName, payload)
}
\arguments{
\item{thingName}{[required] The name of the thing.}

\item{shadowName}{The name of the shadow.}

\item{payload}{[required] The state information, in JSON format.}
}
\description{
Updates the shadow for the specified thing.
}
\details{
For more information, see
\href{https://docs.aws.amazon.com/iot/latest/developerguide/device-shadow-rest-api.html#API_UpdateThingShadow}{UpdateThingShadow}
in the AWS IoT Developer Guide.
}
\section{Request syntax}{
\preformatted{svc$update_thing_shadow(
  thingName = "string",
  shadowName = "string",
  payload = raw
)
}
}

\keyword{internal}
