% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_create_suite_definition}
\alias{iotdeviceadvisor_create_suite_definition}
\title{Creates a Device Advisor test suite}
\usage{
iotdeviceadvisor_create_suite_definition(suiteDefinitionConfiguration,
  tags)
}
\arguments{
\item{suiteDefinitionConfiguration}{Creates a Device Advisor test suite with suite definition configuration.}

\item{tags}{The tags to be attached to the suite definition.}
}
\description{
Creates a Device Advisor test suite.
}
\section{Request syntax}{
\preformatted{svc$create_suite_definition(
  suiteDefinitionConfiguration = list(
    suiteDefinitionName = "string",
    devices = list(
      list(
        thingArn = "string",
        certificateArn = "string"
      )
    ),
    intendedForQualification = TRUE|FALSE,
    rootGroup = "string",
    devicePermissionRoleArn = "string"
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
