% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_delete_suite_definition}
\alias{iotdeviceadvisor_delete_suite_definition}
\title{Deletes a Device Advisor test suite}
\usage{
iotdeviceadvisor_delete_suite_definition(suiteDefinitionId)
}
\arguments{
\item{suiteDefinitionId}{[required] Deletes a Device Advisor test suite with defined suite Id.}
}
\description{
Deletes a Device Advisor test suite.
}
\section{Request syntax}{
\preformatted{svc$delete_suite_definition(
  suiteDefinitionId = "string"
)
}
}

\keyword{internal}
