% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_get_suite_definition}
\alias{iotdeviceadvisor_get_suite_definition}
\title{Gets information about a Device Advisor test suite}
\usage{
iotdeviceadvisor_get_suite_definition(suiteDefinitionId,
  suiteDefinitionVersion)
}
\arguments{
\item{suiteDefinitionId}{[required] Requests suite definition Id with GetSuiteDefinition API call.}

\item{suiteDefinitionVersion}{Requests the suite definition version of a test suite.}
}
\description{
Gets information about a Device Advisor test suite.
}
\section{Request syntax}{
\preformatted{svc$get_suite_definition(
  suiteDefinitionId = "string",
  suiteDefinitionVersion = "string"
)
}
}

\keyword{internal}
