% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_get_suite_run}
\alias{iotdeviceadvisor_get_suite_run}
\title{Gets information about a Device Advisor test suite run}
\usage{
iotdeviceadvisor_get_suite_run(suiteDefinitionId, suiteRunId)
}
\arguments{
\item{suiteDefinitionId}{[required] Requests the information about Device Advisor test suite run based on
suite definition Id.}

\item{suiteRunId}{[required] Requests the information about Device Advisor test suite run based on
suite run Id.}
}
\description{
Gets information about a Device Advisor test suite run.
}
\section{Request syntax}{
\preformatted{svc$get_suite_run(
  suiteDefinitionId = "string",
  suiteRunId = "string"
)
}
}

\keyword{internal}
