% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_get_suite_run_report}
\alias{iotdeviceadvisor_get_suite_run_report}
\title{Gets a report download link for a successful Device Advisor qualifying
test suite run}
\usage{
iotdeviceadvisor_get_suite_run_report(suiteDefinitionId, suiteRunId)
}
\arguments{
\item{suiteDefinitionId}{[required] Device Advisor suite definition Id.}

\item{suiteRunId}{[required] Device Advisor suite run Id.}
}
\description{
Gets a report download link for a successful Device Advisor qualifying
test suite run.
}
\section{Request syntax}{
\preformatted{svc$get_suite_run_report(
  suiteDefinitionId = "string",
  suiteRunId = "string"
)
}
}

\keyword{internal}
