% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_list_suite_runs}
\alias{iotdeviceadvisor_list_suite_runs}
\title{Lists the runs of the specified Device Advisor test suite}
\usage{
iotdeviceadvisor_list_suite_runs(suiteDefinitionId,
  suiteDefinitionVersion, maxResults, nextToken)
}
\arguments{
\item{suiteDefinitionId}{Lists the runs of the specified Device Advisor test suite based on suite
definition Id.}

\item{suiteDefinitionVersion}{Lists the runs of the specified Device Advisor test suite based on suite
definition version.}

\item{maxResults}{MaxResults for list suite run API request.}

\item{nextToken}{Next pagination token for list suite run request.}
}
\description{
Lists the runs of the specified Device Advisor test suite. You can list
all runs of the test suite, or the runs of a specific version of the
test suite.
}
\section{Request syntax}{
\preformatted{svc$list_suite_runs(
  suiteDefinitionId = "string",
  suiteDefinitionVersion = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
