% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_list_tags_for_resource}
\alias{iotdeviceadvisor_list_tags_for_resource}
\title{Lists the tags attached to an IoT Device Advisor resource}
\usage{
iotdeviceadvisor_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the IoT Device Advisor resource.}
}
\description{
Lists the tags attached to an IoT Device Advisor resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
