% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_list_test_cases}
\alias{iotdeviceadvisor_list_test_cases}
\title{Lists all the test cases in the test suite}
\usage{
iotdeviceadvisor_list_test_cases(intendedForQualification, maxResults,
  nextToken)
}
\arguments{
\item{intendedForQualification}{Lists all the qualification test cases in the test suite.}

\item{maxResults}{Requests the test cases max results.}

\item{nextToken}{Requests the test cases next token.}
}
\description{
Lists all the test cases in the test suite.
}
\section{Request syntax}{
\preformatted{svc$list_test_cases(
  intendedForQualification = TRUE|FALSE,
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
