% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_start_suite_run}
\alias{iotdeviceadvisor_start_suite_run}
\title{Starts a Device Advisor test suite run}
\usage{
iotdeviceadvisor_start_suite_run(suiteDefinitionId,
  suiteDefinitionVersion, suiteRunConfiguration, tags)
}
\arguments{
\item{suiteDefinitionId}{[required] Request to start suite run based on suite definition Id.}

\item{suiteDefinitionVersion}{Request to start suite run based on suite definition version.}

\item{suiteRunConfiguration}{Request to start suite run based on suite configuration.}

\item{tags}{The tags to be attached to the suite run.}
}
\description{
Starts a Device Advisor test suite run.
}
\section{Request syntax}{
\preformatted{svc$start_suite_run(
  suiteDefinitionId = "string",
  suiteDefinitionVersion = "string",
  suiteRunConfiguration = list(
    primaryDevice = list(
      thingArn = "string",
      certificateArn = "string"
    ),
    secondaryDevice = list(
      thingArn = "string",
      certificateArn = "string"
    ),
    selectedTestList = list(
      "string"
    )
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
