% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_tag_resource}
\alias{iotdeviceadvisor_tag_resource}
\title{Adds to and modifies existing tags of an IoT Device Advisor resource}
\usage{
iotdeviceadvisor_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The resource ARN of an IoT Device Advisor resource.}

\item{tags}{[required] The tags to be attached to the IoT Device Advisor resource.}
}
\description{
Adds to and modifies existing tags of an IoT Device Advisor resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
