% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_untag_resource}
\alias{iotdeviceadvisor_untag_resource}
\title{Removes tags from an IoT Device Advisor resource}
\usage{
iotdeviceadvisor_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The resource ARN of an IoT Device Advisor resource.}

\item{tagKeys}{[required] List of tag keys to remove from the IoT Device Advisor resource.}
}
\description{
Removes tags from an IoT Device Advisor resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
