% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdeviceadvisor_operations.R
\name{iotdeviceadvisor_update_suite_definition}
\alias{iotdeviceadvisor_update_suite_definition}
\title{Updates a Device Advisor test suite}
\usage{
iotdeviceadvisor_update_suite_definition(suiteDefinitionId,
  suiteDefinitionConfiguration)
}
\arguments{
\item{suiteDefinitionId}{[required] Updates a Device Advisor test suite with suite definition id.}

\item{suiteDefinitionConfiguration}{Updates a Device Advisor test suite with suite definition configuration.}
}
\description{
Updates a Device Advisor test suite.
}
\section{Request syntax}{
\preformatted{svc$update_suite_definition(
  suiteDefinitionId = "string",
  suiteDefinitionConfiguration = list(
    suiteDefinitionName = "string",
    devices = list(
      list(
        thingArn = "string",
        certificateArn = "string"
      )
    ),
    intendedForQualification = TRUE|FALSE,
    rootGroup = "string",
    devicePermissionRoleArn = "string"
  )
)
}
}

\keyword{internal}
