% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_create_input}
\alias{iotevents_create_input}
\title{Creates an input}
\usage{
iotevents_create_input(inputName, inputDescription, inputDefinition,
  tags)
}
\arguments{
\item{inputName}{[required] The name you want to give to the input.}

\item{inputDescription}{A brief description of the input.}

\item{inputDefinition}{[required] The definition of the input.}

\item{tags}{Metadata which can be used to manage the input.}
}
\description{
Creates an input.
}
\section{Request syntax}{
\preformatted{svc$create_input(
  inputName = "string",
  inputDescription = "string",
  inputDefinition = list(
    attributes = list(
      list(
        jsonPath = "string"
      )
    )
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
