% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_delete_detector_model}
\alias{iotevents_delete_detector_model}
\title{Deletes a detector model}
\usage{
iotevents_delete_detector_model(detectorModelName)
}
\arguments{
\item{detectorModelName}{[required] The name of the detector model to be deleted.}
}
\description{
Deletes a detector model. Any active instances of the detector model are
also deleted.
}
\section{Request syntax}{
\preformatted{svc$delete_detector_model(
  detectorModelName = "string"
)
}
}

\keyword{internal}
