% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_delete_input}
\alias{iotevents_delete_input}
\title{Deletes an input}
\usage{
iotevents_delete_input(inputName)
}
\arguments{
\item{inputName}{[required] The name of the input to be deleted.}
}
\description{
Deletes an input.
}
\section{Request syntax}{
\preformatted{svc$delete_input(
  inputName = "string"
)
}
}

\keyword{internal}
