% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_describe_detector_model}
\alias{iotevents_describe_detector_model}
\title{Describes a detector model}
\usage{
iotevents_describe_detector_model(detectorModelName,
  detectorModelVersion)
}
\arguments{
\item{detectorModelName}{[required] The name of the detector model.}

\item{detectorModelVersion}{The version of the detector model.}
}
\description{
Describes a detector model. If the \code{version} parameter is not specified,
information about the latest version is returned.
}
\section{Request syntax}{
\preformatted{svc$describe_detector_model(
  detectorModelName = "string",
  detectorModelVersion = "string"
)
}
}

\keyword{internal}
