% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_describe_input}
\alias{iotevents_describe_input}
\title{Describes an input}
\usage{
iotevents_describe_input(inputName)
}
\arguments{
\item{inputName}{[required] The name of the input.}
}
\description{
Describes an input.
}
\section{Request syntax}{
\preformatted{svc$describe_input(
  inputName = "string"
)
}
}

\keyword{internal}
