% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_describe_logging_options}
\alias{iotevents_describe_logging_options}
\title{Retrieves the current settings of the AWS IoT Events logging options}
\usage{
iotevents_describe_logging_options()
}
\description{
Retrieves the current settings of the AWS IoT Events logging options.
}
\section{Request syntax}{
\preformatted{svc$describe_logging_options()
}
}

\keyword{internal}
