% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_list_detector_model_versions}
\alias{iotevents_list_detector_model_versions}
\title{Lists all the versions of a detector model}
\usage{
iotevents_list_detector_model_versions(detectorModelName, nextToken,
  maxResults)
}
\arguments{
\item{detectorModelName}{[required] The name of the detector model whose versions are returned.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists all the versions of a detector model. Only the metadata associated
with each detector model version is returned.
}
\section{Request syntax}{
\preformatted{svc$list_detector_model_versions(
  detectorModelName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
