% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_list_detector_models}
\alias{iotevents_list_detector_models}
\title{Lists the detector models you have created}
\usage{
iotevents_list_detector_models(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the detector models you have created. Only the metadata associated
with each detector model is returned.
}
\section{Request syntax}{
\preformatted{svc$list_detector_models(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
