% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_list_inputs}
\alias{iotevents_list_inputs}
\title{Lists the inputs you have created}
\usage{
iotevents_list_inputs(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the inputs you have created.
}
\section{Request syntax}{
\preformatted{svc$list_inputs(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
