% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_list_tags_for_resource}
\alias{iotevents_list_tags_for_resource}
\title{Lists the tags (metadata) you have assigned to the resource}
\usage{
iotevents_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}
}
\description{
Lists the tags (metadata) you have assigned to the resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
