% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_put_logging_options}
\alias{iotevents_put_logging_options}
\title{Sets or updates the AWS IoT Events logging options}
\usage{
iotevents_put_logging_options(loggingOptions)
}
\arguments{
\item{loggingOptions}{[required] The new values of the AWS IoT Events logging options.}
}
\description{
Sets or updates the AWS IoT Events logging options.
}
\details{
Note that if you update the value of any \code{loggingOptions} field, it
takes up to one minute for the change to take effect. Also, if you
change the policy attached to the role you specified in the roleArn
field (for example, to correct an invalid policy) it takes up to five
minutes for that change to take effect.
}
\section{Request syntax}{
\preformatted{svc$put_logging_options(
  loggingOptions = list(
    roleArn = "string",
    level = "ERROR"|"INFO"|"DEBUG",
    enabled = TRUE|FALSE,
    detectorDebugOptions = list(
      list(
        detectorModelName = "string",
        keyValue = "string"
      )
    )
  )
)
}
}

\keyword{internal}
