% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_tag_resource}
\alias{iotevents_tag_resource}
\title{Add to or modifies the tags of the given resource}
\usage{
iotevents_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tags}{[required] The new or modified tags for the resource.}
}
\description{
Add to or modifies the tags of the given resource. Tags are metadata
which can be used to manage a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
