% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_update_detector_model}
\alias{iotevents_update_detector_model}
\title{Updates a detector model}
\usage{
iotevents_update_detector_model(detectorModelName,
  detectorModelDefinition, detectorModelDescription, roleArn)
}
\arguments{
\item{detectorModelName}{[required] The name of the detector model to be updated.}

\item{detectorModelDefinition}{[required] Information that defines how a detector operates.}

\item{detectorModelDescription}{A brief description of the detector model.}

\item{roleArn}{[required] The ARN of the role that grants permission to AWS IoT Events to perform
its operations.}
}
\description{
Updates a detector model. Detectors (instances) spawned by the previous
version will be deleted and re-created as new inputs arrive.
}
\section{Request syntax}{
\preformatted{svc$update_detector_model(
  detectorModelName = "string",
  detectorModelDefinition = list(
    states = list(
      list(
        stateName = "string",
        onInput = list(
          events = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string"
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string"
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  )
                )
              )
            )
          ),
          transitionEvents = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string"
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string"
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  )
                )
              ),
              nextState = "string"
            )
          )
        ),
        onEnter = list(
          events = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string"
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string"
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  )
                )
              )
            )
          )
        ),
        onExit = list(
          events = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string"
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string"
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  )
                )
              )
            )
          )
        )
      )
    ),
    initialStateName = "string"
  ),
  detectorModelDescription = "string",
  roleArn = "string"
)
}
}

\keyword{internal}
