% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_update_input}
\alias{iotevents_update_input}
\title{Updates an input}
\usage{
iotevents_update_input(inputName, inputDescription, inputDefinition)
}
\arguments{
\item{inputName}{[required] The name of the input you want to update.}

\item{inputDescription}{A brief description of the input.}

\item{inputDefinition}{[required] The definition of the input.}
}
\description{
Updates an input.
}
\section{Request syntax}{
\preformatted{svc$update_input(
  inputName = "string",
  inputDescription = "string",
  inputDefinition = list(
    attributes = list(
      list(
        jsonPath = "string"
      )
    )
  )
)
}
}

\keyword{internal}
