% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioteventsdata_operations.R
\name{ioteventsdata_batch_put_message}
\alias{ioteventsdata_batch_put_message}
\title{Sends a set of messages to the AWS IoT Events system}
\usage{
ioteventsdata_batch_put_message(messages)
}
\arguments{
\item{messages}{[required] The list of messages to send. Each message has format:
\code{'\{ "messageId": "string", "inputName": "string", "payload": "string"\}'}.}
}
\description{
Sends a set of messages to the AWS IoT Events system. Each message
payload will be transformed into the input you specify (\code{inputName}) and
ingested into any detectors that monitor that input. If multiple
messages are sent, the order in which the messages are processed is not
guaranteed\--you must send messages one at a time and wait for a
successful response to guarantee ordering.
}
\section{Request syntax}{
\preformatted{svc$batch_put_message(
  messages = list(
    list(
      messageId = "string",
      inputName = "string",
      payload = raw
    )
  )
)
}
}

\keyword{internal}
