% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioteventsdata_operations.R
\name{ioteventsdata_batch_update_detector}
\alias{ioteventsdata_batch_update_detector}
\title{Updates the state, variable values, and timer settings of one or more
detectors (instances) of a specified detector model}
\usage{
ioteventsdata_batch_update_detector(detectors)
}
\arguments{
\item{detectors}{[required] The list of detectors (instances) to be updated, along with the values
to be updated.}
}
\description{
Updates the state, variable values, and timer settings of one or more
detectors (instances) of a specified detector model.
}
\section{Request syntax}{
\preformatted{svc$batch_update_detector(
  detectors = list(
    list(
      messageId = "string",
      detectorModelName = "string",
      keyValue = "string",
      state = list(
        stateName = "string",
        variables = list(
          list(
            name = "string",
            value = "string"
          )
        ),
        timers = list(
          list(
            name = "string",
            seconds = 123
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
