% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioteventsdata_operations.R
\name{ioteventsdata_describe_detector}
\alias{ioteventsdata_describe_detector}
\title{Returns information about the specified detector (instance)}
\usage{
ioteventsdata_describe_detector(detectorModelName, keyValue)
}
\arguments{
\item{detectorModelName}{[required] The name of the detector model whose detectors (instances) you want
information about.}

\item{keyValue}{A filter used to limit results to detectors (instances) created because
of the given key ID.}
}
\description{
Returns information about the specified detector (instance).
}
\section{Request syntax}{
\preformatted{svc$describe_detector(
  detectorModelName = "string",
  keyValue = "string"
)
}
}

\keyword{internal}
