% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioteventsdata_operations.R
\name{ioteventsdata_list_detectors}
\alias{ioteventsdata_list_detectors}
\title{Lists detectors (the instances of a detector model)}
\usage{
ioteventsdata_list_detectors(detectorModelName, stateName, nextToken,
  maxResults)
}
\arguments{
\item{detectorModelName}{[required] The name of the detector model whose detectors (instances) are listed.}

\item{stateName}{A filter that limits results to those detectors (instances) in the given
state.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists detectors (the instances of a detector model).
}
\section{Request syntax}{
\preformatted{svc$list_detectors(
  detectorModelName = "string",
  stateName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
