% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotfleethub_service.R
\name{iotfleethub}
\alias{iotfleethub}
\title{AWS IoT Fleet Hub}
\usage{
iotfleethub(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
With Fleet Hub for AWS IoT Device Management you can build stand-alone
web applications for monitoring the health of your device fleets.

Fleet Hub for AWS IoT Device Management is in public preview and is
subject to change.
}
\section{Service syntax}{
\preformatted{svc <- iotfleethub(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=iotfleethub_create_application]{create_application} \tab Creates a Fleet Hub for AWS IoT Device Management web application\cr
\link[=iotfleethub_delete_application]{delete_application} \tab Deletes a Fleet Hub for AWS IoT Device Management web application\cr
\link[=iotfleethub_describe_application]{describe_application} \tab Gets information about a Fleet Hub for AWS IoT Device Management web application\cr
\link[=iotfleethub_list_applications]{list_applications} \tab Gets a list of Fleet Hub for AWS IoT Device Management web applications for the current account\cr
\link[=iotfleethub_list_tags_for_resource]{list_tags_for_resource} \tab Lists the tags for the specified resource\cr
\link[=iotfleethub_tag_resource]{tag_resource} \tab Adds to or modifies the tags of the specified resource\cr
\link[=iotfleethub_untag_resource]{untag_resource} \tab Removes the specified tags (metadata) from the resource\cr
\link[=iotfleethub_update_application]{update_application} \tab Updates information about a Fleet Hub for a AWS IoT Device Management web application
}
}

\examples{
\dontrun{
svc <- iotfleethub()
svc$create_application(
  Foo = 123
)
}

}
