% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotfleethub_operations.R
\name{iotfleethub_delete_application}
\alias{iotfleethub_delete_application}
\title{Deletes a Fleet Hub for AWS IoT Device Management web application}
\usage{
iotfleethub_delete_application(applicationId, clientToken)
}
\arguments{
\item{applicationId}{[required] The unique Id of the web application.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don't reuse this client token if a new
idempotent request is required.}
}
\description{
Deletes a Fleet Hub for AWS IoT Device Management web application.
}
\details{
Fleet Hub for AWS IoT Device Management is in public preview and is
subject to change.
}
\section{Request syntax}{
\preformatted{svc$delete_application(
  applicationId = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
