% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotfleethub_operations.R
\name{iotfleethub_describe_application}
\alias{iotfleethub_describe_application}
\title{Gets information about a Fleet Hub for AWS IoT Device Management web
application}
\usage{
iotfleethub_describe_application(applicationId)
}
\arguments{
\item{applicationId}{[required] The unique Id of the web application.}
}
\description{
Gets information about a Fleet Hub for AWS IoT Device Management web
application.
}
\details{
Fleet Hub for AWS IoT Device Management is in public preview and is
subject to change.
}
\section{Request syntax}{
\preformatted{svc$describe_application(
  applicationId = "string"
)
}
}

\keyword{internal}
