% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotfleethub_operations.R
\name{iotfleethub_list_applications}
\alias{iotfleethub_list_applications}
\title{Gets a list of Fleet Hub for AWS IoT Device Management web applications
for the current account}
\usage{
iotfleethub_list_applications(nextToken)
}
\arguments{
\item{nextToken}{A token used to get the next set of results.}
}
\description{
Gets a list of Fleet Hub for AWS IoT Device Management web applications
for the current account.

Fleet Hub for AWS IoT Device Management is in public preview and is
subject to change.
}
\section{Request syntax}{
\preformatted{svc$list_applications(
  nextToken = "string"
)
}
}

\keyword{internal}
