% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotfleethub_operations.R
\name{iotfleethub_tag_resource}
\alias{iotfleethub_tag_resource}
\title{Adds to or modifies the tags of the specified resource}
\usage{
iotfleethub_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tags}{[required] The new or modified tags for the resource.}
}
\description{
Adds to or modifies the tags of the specified resource. Tags are
metadata which can be used to manage a resource.
}
\details{
Fleet Hub for AWS IoT Device Management is in public preview and is
subject to change.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
