% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotfleethub_operations.R
\name{iotfleethub_untag_resource}
\alias{iotfleethub_untag_resource}
\title{Removes the specified tags (metadata) from the resource}
\usage{
iotfleethub_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tagKeys}{[required] A list of the keys of the tags to be removed from the resource.}
}
\description{
Removes the specified tags (metadata) from the resource.
}
\details{
Fleet Hub for AWS IoT Device Management is in public preview and is
subject to change.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
