% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsecuretunneling_service.R
\name{iotsecuretunneling}
\alias{iotsecuretunneling}
\title{AWS IoT Secure Tunneling}
\usage{
iotsecuretunneling(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
AWS IoT Secure Tunnling enables you to create remote connections to
devices deployed in the field.

For more information about how AWS IoT Secure Tunneling works, see the
User Guide.
}
\section{Service syntax}{
\preformatted{svc <- iotsecuretunneling(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=iotsecuretunneling_close_tunnel]{close_tunnel} \tab Closes a tunnel identified by the unique tunnel id \cr
\link[=iotsecuretunneling_describe_tunnel]{describe_tunnel} \tab Gets information about a tunnel identified by the unique tunnel id \cr
\link[=iotsecuretunneling_list_tags_for_resource]{list_tags_for_resource} \tab Lists the tags for the specified resource \cr
\link[=iotsecuretunneling_list_tunnels]{list_tunnels} \tab List all tunnels for an AWS account \cr
\link[=iotsecuretunneling_open_tunnel]{open_tunnel} \tab Creates a new tunnel, and returns two client access tokens for clients to use to connect to the AWS IoT Secure Tunneling proxy server\cr
\link[=iotsecuretunneling_tag_resource]{tag_resource} \tab A resource tag \cr
\link[=iotsecuretunneling_untag_resource]{untag_resource} \tab Removes a tag from a resource
}
}

\examples{
\dontrun{svc <- iotsecuretunneling()
svc$close_tunnel(
  Foo = 123
)}

}
