% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsecuretunneling_operations.R
\name{iotsecuretunneling_close_tunnel}
\alias{iotsecuretunneling_close_tunnel}
\title{Closes a tunnel identified by the unique tunnel id}
\usage{
iotsecuretunneling_close_tunnel(tunnelId, delete)
}
\arguments{
\item{tunnelId}{[required] The ID of the tunnel to close.}

\item{delete}{When set to true, AWS IoT Secure Tunneling deletes the tunnel data
immediately.}
}
\description{
Closes a tunnel identified by the unique tunnel id. When a
\code{\link[=iotsecuretunneling_close_tunnel]{close_tunnel}} request is received,
we close the WebSocket connections between the client and proxy server
so no data can be transmitted.
}
\section{Request syntax}{
\preformatted{svc$close_tunnel(
  tunnelId = "string",
  delete = TRUE|FALSE
)
}
}

\keyword{internal}
