% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsecuretunneling_operations.R
\name{iotsecuretunneling_describe_tunnel}
\alias{iotsecuretunneling_describe_tunnel}
\title{Gets information about a tunnel identified by the unique tunnel id}
\usage{
iotsecuretunneling_describe_tunnel(tunnelId)
}
\arguments{
\item{tunnelId}{[required] The tunnel to describe.}
}
\description{
Gets information about a tunnel identified by the unique tunnel id.
}
\section{Request syntax}{
\preformatted{svc$describe_tunnel(
  tunnelId = "string"
)
}
}

\keyword{internal}
