% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsecuretunneling_operations.R
\name{iotsecuretunneling_list_tags_for_resource}
\alias{iotsecuretunneling_list_tags_for_resource}
\title{Lists the tags for the specified resource}
\usage{
iotsecuretunneling_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The resource ARN.}
}
\description{
Lists the tags for the specified resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
