% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsecuretunneling_operations.R
\name{iotsecuretunneling_list_tunnels}
\alias{iotsecuretunneling_list_tunnels}
\title{List all tunnels for an AWS account}
\usage{
iotsecuretunneling_list_tunnels(thingName, maxResults, nextToken)
}
\arguments{
\item{thingName}{The name of the IoT thing associated with the destination device.}

\item{maxResults}{The maximum number of results to return at once.}

\item{nextToken}{A token to retrieve the next set of results.}
}
\description{
List all tunnels for an AWS account. Tunnels are listed by creation time
in descending order, newer tunnels will be listed before older tunnels.
}
\section{Request syntax}{
\preformatted{svc$list_tunnels(
  thingName = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
