% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsecuretunneling_operations.R
\name{iotsecuretunneling_open_tunnel}
\alias{iotsecuretunneling_open_tunnel}
\title{Creates a new tunnel, and returns two client access tokens for clients
to use to connect to the AWS IoT Secure Tunneling proxy server}
\usage{
iotsecuretunneling_open_tunnel(description, tags, destinationConfig,
  timeoutConfig)
}
\arguments{
\item{description}{A short text description of the tunnel.}

\item{tags}{A collection of tag metadata.}

\item{destinationConfig}{The destination configuration for the OpenTunnel request.}

\item{timeoutConfig}{Timeout configuration for a tunnel.}
}
\description{
Creates a new tunnel, and returns two client access tokens for clients
to use to connect to the AWS IoT Secure Tunneling proxy server.
}
\section{Request syntax}{
\preformatted{svc$open_tunnel(
  description = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  ),
  destinationConfig = list(
    thingName = "string",
    services = list(
      "string"
    )
  ),
  timeoutConfig = list(
    maxLifetimeTimeoutMinutes = 123
  )
)
}
}

\keyword{internal}
