% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsecuretunneling_operations.R
\name{iotsecuretunneling_untag_resource}
\alias{iotsecuretunneling_untag_resource}
\title{Removes a tag from a resource}
\usage{
iotsecuretunneling_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The resource ARN.}

\item{tagKeys}{[required] The keys of the tags to remove.}
}
\description{
Removes a tag from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
