% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_service.R
\name{iotsitewise}
\alias{iotsitewise}
\title{AWS IoT SiteWise}
\usage{
iotsitewise(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Welcome to the AWS IoT SiteWise API Reference. AWS IoT SiteWise is an
AWS service that connects \href{https://en.wikipedia.org/wiki/Internet_of_things#Industrial_applications}{Industrial Internet of Things (IIoT)}
devices to the power of the AWS Cloud. For more information, see the
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/}{AWS IoT SiteWise User Guide}. For
information about AWS IoT SiteWise quotas, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/quotas.html}{Quotas}
in the \emph{AWS IoT SiteWise User Guide}.
}
\section{Service syntax}{
\preformatted{svc <- iotsitewise(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=iotsitewise_associate_assets]{associate_assets} \tab Associates a child asset with the given parent asset through a hierarchy defined in the parent asset's model\cr
\link[=iotsitewise_batch_associate_project_assets]{batch_associate_project_assets} \tab Associates a group (batch) of assets with an AWS IoT SiteWise Monitor project\cr
\link[=iotsitewise_batch_disassociate_project_assets]{batch_disassociate_project_assets} \tab Disassociates a group (batch) of assets from an AWS IoT SiteWise Monitor project\cr
\link[=iotsitewise_batch_put_asset_property_value]{batch_put_asset_property_value} \tab Sends a list of asset property values to AWS IoT SiteWise\cr
\link[=iotsitewise_create_access_policy]{create_access_policy} \tab Creates an access policy that grants the specified AWS Single Sign-On user or group access to the specified AWS IoT SiteWise Monitor portal or project resource\cr
\link[=iotsitewise_create_asset]{create_asset} \tab Creates an asset from an existing asset model\cr
\link[=iotsitewise_create_asset_model]{create_asset_model} \tab Creates an asset model from specified property and hierarchy definitions\cr
\link[=iotsitewise_create_dashboard]{create_dashboard} \tab Creates a dashboard in an AWS IoT SiteWise Monitor project\cr
\link[=iotsitewise_create_gateway]{create_gateway} \tab Creates a gateway, which is a virtual or edge device that delivers industrial data streams from local servers to AWS IoT SiteWise\cr
\link[=iotsitewise_create_portal]{create_portal} \tab Creates a portal, which can contain projects and dashboards\cr
\link[=iotsitewise_create_project]{create_project} \tab Creates a project in the specified portal\cr
\link[=iotsitewise_delete_access_policy]{delete_access_policy} \tab Deletes an access policy that grants the specified AWS Single Sign-On identity access to the specified AWS IoT SiteWise Monitor resource\cr
\link[=iotsitewise_delete_asset]{delete_asset} \tab Deletes an asset\cr
\link[=iotsitewise_delete_asset_model]{delete_asset_model} \tab Deletes an asset model\cr
\link[=iotsitewise_delete_dashboard]{delete_dashboard} \tab Deletes a dashboard from AWS IoT SiteWise Monitor\cr
\link[=iotsitewise_delete_gateway]{delete_gateway} \tab Deletes a gateway from AWS IoT SiteWise\cr
\link[=iotsitewise_delete_portal]{delete_portal} \tab Deletes a portal from AWS IoT SiteWise Monitor\cr
\link[=iotsitewise_delete_project]{delete_project} \tab Deletes a project from AWS IoT SiteWise Monitor\cr
\link[=iotsitewise_describe_access_policy]{describe_access_policy} \tab Describes an access policy, which specifies an AWS SSO user or group's access to an AWS IoT SiteWise Monitor portal or project\cr
\link[=iotsitewise_describe_asset]{describe_asset} \tab Retrieves information about an asset\cr
\link[=iotsitewise_describe_asset_model]{describe_asset_model} \tab Retrieves information about an asset model\cr
\link[=iotsitewise_describe_asset_property]{describe_asset_property} \tab Retrieves information about an asset's property\cr
\link[=iotsitewise_describe_dashboard]{describe_dashboard} \tab Retrieves information about a dashboard\cr
\link[=iotsitewise_describe_gateway]{describe_gateway} \tab Retrieves information about a gateway\cr
\link[=iotsitewise_describe_gateway_capability_configuration]{describe_gateway_capability_configuration} \tab Retrieves information about a gateway capability configuration\cr
\link[=iotsitewise_describe_logging_options]{describe_logging_options} \tab Retrieves the current AWS IoT SiteWise logging options\cr
\link[=iotsitewise_describe_portal]{describe_portal} \tab Retrieves information about a portal\cr
\link[=iotsitewise_describe_project]{describe_project} \tab Retrieves information about a project\cr
\link[=iotsitewise_disassociate_assets]{disassociate_assets} \tab Disassociates a child asset from the given parent asset through a hierarchy defined in the parent asset's model\cr
\link[=iotsitewise_get_asset_property_aggregates]{get_asset_property_aggregates} \tab Gets aggregated values for an asset property\cr
\link[=iotsitewise_get_asset_property_value]{get_asset_property_value} \tab Gets an asset property's current value\cr
\link[=iotsitewise_get_asset_property_value_history]{get_asset_property_value_history} \tab Gets the history of an asset property's values\cr
\link[=iotsitewise_list_access_policies]{list_access_policies} \tab Retrieves a paginated list of access policies for an AWS SSO identity (a user or group) or an AWS IoT SiteWise Monitor resource (a portal or project)\cr
\link[=iotsitewise_list_asset_models]{list_asset_models} \tab Retrieves a paginated list of summaries of all asset models\cr
\link[=iotsitewise_list_assets]{list_assets} \tab Retrieves a paginated list of asset summaries\cr
\link[=iotsitewise_list_associated_assets]{list_associated_assets} \tab Retrieves a paginated list of the assets associated to a parent asset (assetId) by a given hierarchy (hierarchyId)\cr
\link[=iotsitewise_list_dashboards]{list_dashboards} \tab Retrieves a paginated list of dashboards for an AWS IoT SiteWise Monitor project\cr
\link[=iotsitewise_list_gateways]{list_gateways} \tab Retrieves a paginated list of gateways\cr
\link[=iotsitewise_list_portals]{list_portals} \tab Retrieves a paginated list of AWS IoT SiteWise Monitor portals\cr
\link[=iotsitewise_list_project_assets]{list_project_assets} \tab Retrieves a paginated list of assets associated with an AWS IoT SiteWise Monitor project\cr
\link[=iotsitewise_list_projects]{list_projects} \tab Retrieves a paginated list of projects for an AWS IoT SiteWise Monitor portal\cr
\link[=iotsitewise_list_tags_for_resource]{list_tags_for_resource} \tab Retrieves the list of tags for an AWS IoT SiteWise resource\cr
\link[=iotsitewise_put_logging_options]{put_logging_options} \tab Sets logging options for AWS IoT SiteWise\cr
\link[=iotsitewise_tag_resource]{tag_resource} \tab Adds tags to an AWS IoT SiteWise resource\cr
\link[=iotsitewise_untag_resource]{untag_resource} \tab Removes a tag from an AWS IoT SiteWise resource\cr
\link[=iotsitewise_update_access_policy]{update_access_policy} \tab Updates an existing access policy that specifies an AWS SSO user or group's access to an AWS IoT SiteWise Monitor portal or project resource\cr
\link[=iotsitewise_update_asset]{update_asset} \tab Updates an asset's name\cr
\link[=iotsitewise_update_asset_model]{update_asset_model} \tab Updates an asset model and all of the assets that were created from the model\cr
\link[=iotsitewise_update_asset_property]{update_asset_property} \tab Updates an asset property's alias and notification state\cr
\link[=iotsitewise_update_dashboard]{update_dashboard} \tab Updates an AWS IoT SiteWise Monitor dashboard\cr
\link[=iotsitewise_update_gateway]{update_gateway} \tab Updates a gateway's name\cr
\link[=iotsitewise_update_gateway_capability_configuration]{update_gateway_capability_configuration} \tab Updates a gateway capability configuration or defines a new capability configuration\cr
\link[=iotsitewise_update_portal]{update_portal} \tab Updates an AWS IoT SiteWise Monitor portal\cr
\link[=iotsitewise_update_project]{update_project} \tab Updates an AWS IoT SiteWise Monitor project
}
}

\examples{
\dontrun{
svc <- iotsitewise()
svc$associate_assets(
  Foo = 123
)
}

}
