% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_batch_associate_project_assets}
\alias{iotsitewise_batch_associate_project_assets}
\title{Associates a group (batch) of assets with an AWS IoT SiteWise Monitor
project}
\usage{
iotsitewise_batch_associate_project_assets(projectId, assetIds,
  clientToken)
}
\arguments{
\item{projectId}{[required] The ID of the project to which to associate the assets.}

\item{assetIds}{[required] The IDs of the assets to be associated to the project.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don\'t reuse this client token if a new
idempotent request is required.}
}
\description{
Associates a group (batch) of assets with an AWS IoT SiteWise Monitor
project.
}
\section{Request syntax}{
\preformatted{svc$batch_associate_project_assets(
  projectId = "string",
  assetIds = list(
    "string"
  ),
  clientToken = "string"
)
}
}

\keyword{internal}
