% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_create_access_policy}
\alias{iotsitewise_create_access_policy}
\title{Creates an access policy that grants the specified AWS Single Sign-On
user or group access to the specified AWS IoT SiteWise Monitor portal or
project resource}
\usage{
iotsitewise_create_access_policy(accessPolicyIdentity,
  accessPolicyResource, accessPolicyPermission, clientToken, tags)
}
\arguments{
\item{accessPolicyIdentity}{[required] The identity for this access policy. Choose either a \code{user} or a \code{group}
but not both.}

\item{accessPolicyResource}{[required] The AWS IoT SiteWise Monitor resource for this access policy. Choose
either \code{portal} or \code{project} but not both.}

\item{accessPolicyPermission}{[required] The permission level for this access policy. Note that a project
\code{ADMINISTRATOR} is also known as a project owner.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don\'t reuse this client token if a new
idempotent request is required.}

\item{tags}{A list of key-value pairs that contain metadata for the access policy.
For more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/tag-resources.html}{Tagging your AWS IoT SiteWise resources}
in the \emph{AWS IoT SiteWise User Guide}.}
}
\description{
Creates an access policy that grants the specified AWS Single Sign-On
user or group access to the specified AWS IoT SiteWise Monitor portal or
project resource.
}
\section{Request syntax}{
\preformatted{svc$create_access_policy(
  accessPolicyIdentity = list(
    user = list(
      id = "string"
    ),
    group = list(
      id = "string"
    )
  ),
  accessPolicyResource = list(
    portal = list(
      id = "string"
    ),
    project = list(
      id = "string"
    )
  ),
  accessPolicyPermission = "ADMINISTRATOR"|"VIEWER",
  clientToken = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
