% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_create_asset}
\alias{iotsitewise_create_asset}
\title{Creates an asset from an existing asset model}
\usage{
iotsitewise_create_asset(assetName, assetModelId, clientToken, tags)
}
\arguments{
\item{assetName}{[required] A unique, friendly name for the asset.}

\item{assetModelId}{[required] The ID of the asset model from which to create the asset.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don't reuse this client token if a new
idempotent request is required.}

\item{tags}{A list of key-value pairs that contain metadata for the asset. For more
information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/tag-resources.html}{Tagging your AWS IoT SiteWise resources}
in the \emph{AWS IoT SiteWise User Guide}.}
}
\description{
Creates an asset from an existing asset model. For more information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/create-assets.html}{Creating assets}
in the \emph{AWS IoT SiteWise User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_asset(
  assetName = "string",
  assetModelId = "string",
  clientToken = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
