% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_create_asset_model}
\alias{iotsitewise_create_asset_model}
\title{Creates an asset model from specified property and hierarchy definitions}
\usage{
iotsitewise_create_asset_model(assetModelName, assetModelDescription,
  assetModelProperties, assetModelHierarchies, clientToken, tags)
}
\arguments{
\item{assetModelName}{[required] A unique, friendly name for the asset model.}

\item{assetModelDescription}{A description for the asset model.}

\item{assetModelProperties}{The property definitions of the asset model. For more information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/asset-properties.html}{Asset Properties}
in the \emph{AWS IoT SiteWise User Guide}.

You can specify up to 200 properties per asset model. For more
information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/quotas.html}{Quotas}
in the \emph{AWS IoT SiteWise User Guide}.}

\item{assetModelHierarchies}{The hierarchy definitions of the asset model. Each hierarchy specifies
an asset model whose assets can be children of any other assets created
from this asset model. For more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/asset-hierarchies.html}{Asset Hierarchies}
in the \emph{AWS IoT SiteWise User Guide}.

You can specify up to 10 hierarchies per asset model. For more
information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/quotas.html}{Quotas}
in the \emph{AWS IoT SiteWise User Guide}.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don\'t reuse this client token if a new
idempotent request is required.}

\item{tags}{A list of key-value pairs that contain metadata for the asset model. For
more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/tag-resources.html}{Tagging your AWS IoT SiteWise resources}
in the \emph{AWS IoT SiteWise User Guide}.}
}
\description{
Creates an asset model from specified property and hierarchy
definitions. You create assets from asset models. With asset models, you
can easily create assets of the same type that have standardized
definitions. Each asset created from a model inherits the asset model\'s
property and hierarchy definitions. For more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/define-models.html}{Defining Asset Models}
in the \emph{AWS IoT SiteWise User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_asset_model(
  assetModelName = "string",
  assetModelDescription = "string",
  assetModelProperties = list(
    list(
      name = "string",
      dataType = "STRING"|"INTEGER"|"DOUBLE"|"BOOLEAN",
      unit = "string",
      type = list(
        attribute = list(
          defaultValue = "string"
        ),
        measurement = list(),
        transform = list(
          expression = "string",
          variables = list(
            list(
              name = "string",
              value = list(
                propertyId = "string",
                hierarchyId = "string"
              )
            )
          )
        ),
        metric = list(
          expression = "string",
          variables = list(
            list(
              name = "string",
              value = list(
                propertyId = "string",
                hierarchyId = "string"
              )
            )
          ),
          window = list(
            tumbling = list(
              interval = "string"
            )
          )
        )
      )
    )
  ),
  assetModelHierarchies = list(
    list(
      name = "string",
      childAssetModelId = "string"
    )
  ),
  clientToken = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
