% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_create_dashboard}
\alias{iotsitewise_create_dashboard}
\title{Creates a dashboard in an AWS IoT SiteWise Monitor project}
\usage{
iotsitewise_create_dashboard(projectId, dashboardName,
  dashboardDescription, dashboardDefinition, clientToken, tags)
}
\arguments{
\item{projectId}{[required] The ID of the project in which to create the dashboard.}

\item{dashboardName}{[required] A friendly name for the dashboard.}

\item{dashboardDescription}{A description for the dashboard.}

\item{dashboardDefinition}{[required] The dashboard definition specified in a JSON literal. For detailed
information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/create-dashboards-using-aws-cli.html}{Creating Dashboards (CLI)}
in the \emph{AWS IoT SiteWise User Guide}.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don\'t reuse this client token if a new
idempotent request is required.}

\item{tags}{A list of key-value pairs that contain metadata for the dashboard. For
more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/tag-resources.html}{Tagging your AWS IoT SiteWise resources}
in the \emph{AWS IoT SiteWise User Guide}.}
}
\description{
Creates a dashboard in an AWS IoT SiteWise Monitor project.
}
\section{Request syntax}{
\preformatted{svc$create_dashboard(
  projectId = "string",
  dashboardName = "string",
  dashboardDescription = "string",
  dashboardDefinition = "string",
  clientToken = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
