% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_create_gateway}
\alias{iotsitewise_create_gateway}
\title{Creates a gateway, which is a virtual or edge device that delivers
industrial data streams from local servers to AWS IoT SiteWise}
\usage{
iotsitewise_create_gateway(gatewayName, gatewayPlatform, tags)
}
\arguments{
\item{gatewayName}{[required] A unique, friendly name for the gateway.}

\item{gatewayPlatform}{[required] The gateway's platform. You can only specify one platform in a gateway.}

\item{tags}{A list of key-value pairs that contain metadata for the gateway. For
more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/tag-resources.html}{Tagging your AWS IoT SiteWise resources}
in the \emph{AWS IoT SiteWise User Guide}.}
}
\description{
Creates a gateway, which is a virtual or edge device that delivers
industrial data streams from local servers to AWS IoT SiteWise. For more
information, see Ingesting data using a gateway in the \emph{AWS IoT SiteWise
User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_gateway(
  gatewayName = "string",
  gatewayPlatform = list(
    greengrass = list(
      groupArn = "string"
    )
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
