% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_create_project}
\alias{iotsitewise_create_project}
\title{Creates a project in the specified portal}
\usage{
iotsitewise_create_project(portalId, projectName, projectDescription,
  clientToken, tags)
}
\arguments{
\item{portalId}{[required] The ID of the portal in which to create the project.}

\item{projectName}{[required] A friendly name for the project.}

\item{projectDescription}{A description for the project.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don\'t reuse this client token if a new
idempotent request is required.}

\item{tags}{A list of key-value pairs that contain metadata for the project. For
more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/tag-resources.html}{Tagging your AWS IoT SiteWise resources}
in the \emph{AWS IoT SiteWise User Guide}.}
}
\description{
Creates a project in the specified portal.
}
\section{Request syntax}{
\preformatted{svc$create_project(
  portalId = "string",
  projectName = "string",
  projectDescription = "string",
  clientToken = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
