% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_delete_dashboard}
\alias{iotsitewise_delete_dashboard}
\title{Deletes a dashboard from AWS IoT SiteWise Monitor}
\usage{
iotsitewise_delete_dashboard(dashboardId, clientToken)
}
\arguments{
\item{dashboardId}{[required] The ID of the dashboard to delete.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don\'t reuse this client token if a new
idempotent request is required.}
}
\description{
Deletes a dashboard from AWS IoT SiteWise Monitor.
}
\section{Request syntax}{
\preformatted{svc$delete_dashboard(
  dashboardId = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
