% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_delete_gateway}
\alias{iotsitewise_delete_gateway}
\title{Deletes a gateway from AWS IoT SiteWise}
\usage{
iotsitewise_delete_gateway(gatewayId)
}
\arguments{
\item{gatewayId}{[required] The ID of the gateway to delete.}
}
\description{
Deletes a gateway from AWS IoT SiteWise. When you delete a gateway, some
of the gateway's files remain in your gateway's file system. For more
information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/data-retention.html}{Data retention}
in the \emph{AWS IoT SiteWise User Guide}.
}
\section{Request syntax}{
\preformatted{svc$delete_gateway(
  gatewayId = "string"
)
}
}

\keyword{internal}
