% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_delete_portal}
\alias{iotsitewise_delete_portal}
\title{Deletes a portal from AWS IoT SiteWise Monitor}
\usage{
iotsitewise_delete_portal(portalId, clientToken)
}
\arguments{
\item{portalId}{[required] The ID of the portal to delete.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don\'t reuse this client token if a new
idempotent request is required.}
}
\description{
Deletes a portal from AWS IoT SiteWise Monitor.
}
\section{Request syntax}{
\preformatted{svc$delete_portal(
  portalId = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
