% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_delete_project}
\alias{iotsitewise_delete_project}
\title{Deletes a project from AWS IoT SiteWise Monitor}
\usage{
iotsitewise_delete_project(projectId, clientToken)
}
\arguments{
\item{projectId}{[required] The ID of the project.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don't reuse this client token if a new
idempotent request is required.}
}
\description{
Deletes a project from AWS IoT SiteWise Monitor.
}
\section{Request syntax}{
\preformatted{svc$delete_project(
  projectId = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
