% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_access_policy}
\alias{iotsitewise_describe_access_policy}
\title{Describes an access policy, which specifies an AWS SSO user or group's
access to an AWS IoT SiteWise Monitor portal or project}
\usage{
iotsitewise_describe_access_policy(accessPolicyId)
}
\arguments{
\item{accessPolicyId}{[required] The ID of the access policy.}
}
\description{
Describes an access policy, which specifies an AWS SSO user or group's
access to an AWS IoT SiteWise Monitor portal or project.
}
\section{Request syntax}{
\preformatted{svc$describe_access_policy(
  accessPolicyId = "string"
)
}
}

\keyword{internal}
