% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_asset}
\alias{iotsitewise_describe_asset}
\title{Retrieves information about an asset}
\usage{
iotsitewise_describe_asset(assetId)
}
\arguments{
\item{assetId}{[required] The ID of the asset.}
}
\description{
Retrieves information about an asset.
}
\section{Request syntax}{
\preformatted{svc$describe_asset(
  assetId = "string"
)
}
}

\keyword{internal}
