% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_asset_model}
\alias{iotsitewise_describe_asset_model}
\title{Retrieves information about an asset model}
\usage{
iotsitewise_describe_asset_model(assetModelId)
}
\arguments{
\item{assetModelId}{[required] The ID of the asset model.}
}
\description{
Retrieves information about an asset model.
}
\section{Request syntax}{
\preformatted{svc$describe_asset_model(
  assetModelId = "string"
)
}
}

\keyword{internal}
