% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_asset_property}
\alias{iotsitewise_describe_asset_property}
\title{Retrieves information about an asset property}
\usage{
iotsitewise_describe_asset_property(assetId, propertyId)
}
\arguments{
\item{assetId}{[required] The ID of the asset.}

\item{propertyId}{[required] The ID of the asset property.}
}
\description{
Retrieves information about an asset property.

When you call this operation for an attribute property, this response
includes the default attribute value that you define in the asset model.
If you update the default value in the model, this operation's response
includes the new default value.

This operation doesn't return the value of the asset property. To get
the value of an asset property, use
\code{\link[=iotsitewise_get_asset_property_value]{get_asset_property_value}}.
}
\section{Request syntax}{
\preformatted{svc$describe_asset_property(
  assetId = "string",
  propertyId = "string"
)
}
}

\keyword{internal}
