% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_dashboard}
\alias{iotsitewise_describe_dashboard}
\title{Retrieves information about a dashboard}
\usage{
iotsitewise_describe_dashboard(dashboardId)
}
\arguments{
\item{dashboardId}{[required] The ID of the dashboard.}
}
\description{
Retrieves information about a dashboard.
}
\section{Request syntax}{
\preformatted{svc$describe_dashboard(
  dashboardId = "string"
)
}
}

\keyword{internal}
