% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_default_encryption_configuration}
\alias{iotsitewise_describe_default_encryption_configuration}
\title{Retrieves information about the default encryption configuration for the
AWS account in the default or specified region}
\usage{
iotsitewise_describe_default_encryption_configuration()
}
\description{
Retrieves information about the default encryption configuration for the
AWS account in the default or specified region. For more information,
see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/key-management.html}{Key management}
in the \emph{AWS IoT SiteWise User Guide}.
}
\section{Request syntax}{
\preformatted{svc$describe_default_encryption_configuration()
}
}

\keyword{internal}
