% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_gateway}
\alias{iotsitewise_describe_gateway}
\title{Retrieves information about a gateway}
\usage{
iotsitewise_describe_gateway(gatewayId)
}
\arguments{
\item{gatewayId}{[required] The ID of the gateway device.}
}
\description{
Retrieves information about a gateway.
}
\section{Request syntax}{
\preformatted{svc$describe_gateway(
  gatewayId = "string"
)
}
}

\keyword{internal}
