% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_gateway_capability_configuration}
\alias{iotsitewise_describe_gateway_capability_configuration}
\title{Retrieves information about a gateway capability configuration}
\usage{
iotsitewise_describe_gateway_capability_configuration(gatewayId,
  capabilityNamespace)
}
\arguments{
\item{gatewayId}{[required] The ID of the gateway that defines the capability configuration.}

\item{capabilityNamespace}{[required] The namespace of the capability configuration. For example, if you
configure OPC-UA sources from the AWS IoT SiteWise console, your OPC-UA
capability configuration has the namespace
\code{iotsitewise:opcuacollector:version}, where \code{version} is a number such
as \code{1}.}
}
\description{
Retrieves information about a gateway capability configuration. Each
gateway capability defines data sources for a gateway. A capability
configuration can contain multiple data source configurations. If you
define OPC-UA sources for a gateway in the AWS IoT SiteWise console, all
of your OPC-UA sources are stored in one capability configuration. To
list all capability configurations for a gateway, use
\code{\link[=iotsitewise_describe_gateway]{describe_gateway}}.
}
\section{Request syntax}{
\preformatted{svc$describe_gateway_capability_configuration(
  gatewayId = "string",
  capabilityNamespace = "string"
)
}
}

\keyword{internal}
