% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_logging_options}
\alias{iotsitewise_describe_logging_options}
\title{Retrieves the current AWS IoT SiteWise logging options}
\usage{
iotsitewise_describe_logging_options()
}
\description{
Retrieves the current AWS IoT SiteWise logging options.
}
\section{Request syntax}{
\preformatted{svc$describe_logging_options()
}
}

\keyword{internal}
