% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_portal}
\alias{iotsitewise_describe_portal}
\title{Retrieves information about a portal}
\usage{
iotsitewise_describe_portal(portalId)
}
\arguments{
\item{portalId}{[required] The ID of the portal.}
}
\description{
Retrieves information about a portal.
}
\section{Request syntax}{
\preformatted{svc$describe_portal(
  portalId = "string"
)
}
}

\keyword{internal}
