% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_project}
\alias{iotsitewise_describe_project}
\title{Retrieves information about a project}
\usage{
iotsitewise_describe_project(projectId)
}
\arguments{
\item{projectId}{[required] The ID of the project.}
}
\description{
Retrieves information about a project.
}
\section{Request syntax}{
\preformatted{svc$describe_project(
  projectId = "string"
)
}
}

\keyword{internal}
