% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_get_asset_property_value}
\alias{iotsitewise_get_asset_property_value}
\title{Gets an asset property's current value}
\usage{
iotsitewise_get_asset_property_value(assetId, propertyId, propertyAlias)
}
\arguments{
\item{assetId}{The ID of the asset.}

\item{propertyId}{The ID of the asset property.}

\item{propertyAlias}{The property alias that identifies the property, such as an OPC-UA
server data stream path (for example,
\verb{/company/windfarm/3/turbine/7/temperature}). For more information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/connect-data-streams.html}{Mapping industrial data streams to asset properties}
in the \emph{AWS IoT SiteWise User Guide}.}
}
\description{
Gets an asset property's current value. For more information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/query-industrial-data.html#current-values}{Querying current values}
in the \emph{AWS IoT SiteWise User Guide}.
}
\details{
To identify an asset property, you must specify one of the following:
\itemize{
\item The \code{assetId} and \code{propertyId} of an asset property.
\item A \code{propertyAlias}, which is a data stream alias (for example,
\verb{/company/windfarm/3/turbine/7/temperature}). To define an asset
property's alias, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/APIReference/API_UpdateAssetProperty.html}{UpdateAssetProperty}.
}
}
\section{Request syntax}{
\preformatted{svc$get_asset_property_value(
  assetId = "string",
  propertyId = "string",
  propertyAlias = "string"
)
}
}

\keyword{internal}
