% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_get_asset_property_value_history}
\alias{iotsitewise_get_asset_property_value_history}
\title{Gets the history of an asset property's values}
\usage{
iotsitewise_get_asset_property_value_history(assetId, propertyId,
  propertyAlias, startDate, endDate, qualities, timeOrdering, nextToken,
  maxResults)
}
\arguments{
\item{assetId}{The ID of the asset.}

\item{propertyId}{The ID of the asset property.}

\item{propertyAlias}{The property alias that identifies the property, such as an OPC-UA
server data stream path (for example,
\verb{/company/windfarm/3/turbine/7/temperature}). For more information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/connect-data-streams.html}{Mapping industrial data streams to asset properties}
in the \emph{AWS IoT SiteWise User Guide}.}

\item{startDate}{The exclusive start of the range from which to query historical data,
expressed in seconds in Unix epoch time.}

\item{endDate}{The inclusive end of the range from which to query historical data,
expressed in seconds in Unix epoch time.}

\item{qualities}{The quality by which to filter asset data.}

\item{timeOrdering}{The chronological sorting order of the requested information.

Default: \code{ASCENDING}}

\item{nextToken}{The token to be used for the next set of paginated results.}

\item{maxResults}{The maximum number of results to be returned per paginated request.

Default: 100}
}
\description{
Gets the history of an asset property's values. For more information,
see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/query-industrial-data.html#historical-values}{Querying historical values}
in the \emph{AWS IoT SiteWise User Guide}.

To identify an asset property, you must specify one of the following:
\itemize{
\item The \code{assetId} and \code{propertyId} of an asset property.
\item A \code{propertyAlias}, which is a data stream alias (for example,
\verb{/company/windfarm/3/turbine/7/temperature}). To define an asset
property's alias, see
\code{\link[=iotsitewise_update_asset_property]{update_asset_property}}.
}
}
\section{Request syntax}{
\preformatted{svc$get_asset_property_value_history(
  assetId = "string",
  propertyId = "string",
  propertyAlias = "string",
  startDate = as.POSIXct(
    "2015-01-01"
  ),
  endDate = as.POSIXct(
    "2015-01-01"
  ),
  qualities = list(
    "GOOD"|"BAD"|"UNCERTAIN"
  ),
  timeOrdering = "ASCENDING"|"DESCENDING",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
