% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_list_access_policies}
\alias{iotsitewise_list_access_policies}
\title{Retrieves a paginated list of access policies for an AWS SSO identity (a
user or group) or an AWS IoT SiteWise Monitor resource (a portal or
project)}
\usage{
iotsitewise_list_access_policies(identityType, identityId, resourceType,
  resourceId, nextToken, maxResults)
}
\arguments{
\item{identityType}{The type of identity (user or group). This parameter is required if you
specify \code{identityId}.}

\item{identityId}{The ID of the identity. This parameter is required if you specify
\code{identityType}.}

\item{resourceType}{The type of resource (portal or project). This parameter is required if
you specify \code{resourceId}.}

\item{resourceId}{The ID of the resource. This parameter is required if you specify
\code{resourceType}.}

\item{nextToken}{The token to be used for the next set of paginated results.}

\item{maxResults}{The maximum number of results to be returned per paginated request.}
}
\description{
Retrieves a paginated list of access policies for an AWS SSO identity (a
user or group) or an AWS IoT SiteWise Monitor resource (a portal or
project).
}
\section{Request syntax}{
\preformatted{svc$list_access_policies(
  identityType = "USER"|"GROUP",
  identityId = "string",
  resourceType = "PORTAL"|"PROJECT",
  resourceId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
