% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_list_asset_models}
\alias{iotsitewise_list_asset_models}
\title{Retrieves a paginated list of summaries of all asset models}
\usage{
iotsitewise_list_asset_models(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token to be used for the next set of paginated results.}

\item{maxResults}{The maximum number of results to be returned per paginated request.}
}
\description{
Retrieves a paginated list of summaries of all asset models.
}
\section{Request syntax}{
\preformatted{svc$list_asset_models(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
