% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_list_asset_relationships}
\alias{iotsitewise_list_asset_relationships}
\title{Retrieves a paginated list of asset relationships for an asset}
\usage{
iotsitewise_list_asset_relationships(assetId, traversalType, nextToken,
  maxResults)
}
\arguments{
\item{assetId}{[required] The ID of the asset.}

\item{traversalType}{[required] The type of traversal to use to identify asset relationships. Choose the
following option:
\itemize{
\item \code{PATH_TO_ROOT} – Identify the asset's parent assets up to the root
asset. The asset that you specify in \code{assetId} is the first result
in the list of \code{assetRelationshipSummaries}, and the root asset is
the last result.
}}

\item{nextToken}{The token to be used for the next set of paginated results.}

\item{maxResults}{The maximum number of results to be returned per paginated request.}
}
\description{
Retrieves a paginated list of asset relationships for an asset. You can
use this operation to identify an asset's root asset and all associated
assets between that asset and its root.
}
\section{Request syntax}{
\preformatted{svc$list_asset_relationships(
  assetId = "string",
  traversalType = "PATH_TO_ROOT",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
