% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_list_assets}
\alias{iotsitewise_list_assets}
\title{Retrieves a paginated list of asset summaries}
\usage{
iotsitewise_list_assets(nextToken, maxResults, assetModelId, filter)
}
\arguments{
\item{nextToken}{The token to be used for the next set of paginated results.}

\item{maxResults}{The maximum number of results to be returned per paginated request.

Default: 50}

\item{assetModelId}{The ID of the asset model by which to filter the list of assets. This
parameter is required if you choose \code{ALL} for \code{filter}.}

\item{filter}{The filter for the requested list of assets. Choose one of the following
options:
\itemize{
\item \code{ALL} – The list includes all assets for a given asset model ID. The
\code{assetModelId} parameter is required if you filter by \code{ALL}.
\item \code{TOP_LEVEL} – The list includes only top-level assets in the asset
hierarchy tree.
}

Default: \code{ALL}}
}
\description{
Retrieves a paginated list of asset summaries.

You can use this operation to do the following:
\itemize{
\item List assets based on a specific asset model.
\item List top-level assets.
}

You can't use this operation to list all assets. To retrieve summaries
for all of your assets, use
\code{\link[=iotsitewise_list_asset_models]{list_asset_models}} to get all of your
asset model IDs. Then, use ListAssets to get all assets for each asset
model.
}
\section{Request syntax}{
\preformatted{svc$list_assets(
  nextToken = "string",
  maxResults = 123,
  assetModelId = "string",
  filter = "ALL"|"TOP_LEVEL"
)
}
}

\keyword{internal}
