% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_list_portals}
\alias{iotsitewise_list_portals}
\title{Retrieves a paginated list of AWS IoT SiteWise Monitor portals}
\usage{
iotsitewise_list_portals(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token to be used for the next set of paginated results.}

\item{maxResults}{The maximum number of results to be returned per paginated request.}
}
\description{
Retrieves a paginated list of AWS IoT SiteWise Monitor portals.
}
\section{Request syntax}{
\preformatted{svc$list_portals(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
