% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_list_projects}
\alias{iotsitewise_list_projects}
\title{Retrieves a paginated list of projects for an AWS IoT SiteWise Monitor
portal}
\usage{
iotsitewise_list_projects(portalId, nextToken, maxResults)
}
\arguments{
\item{portalId}{[required] The ID of the portal.}

\item{nextToken}{The token to be used for the next set of paginated results.}

\item{maxResults}{The maximum number of results to be returned per paginated request.}
}
\description{
Retrieves a paginated list of projects for an AWS IoT SiteWise Monitor
portal.
}
\section{Request syntax}{
\preformatted{svc$list_projects(
  portalId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
