% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_list_tags_for_resource}
\alias{iotsitewise_list_tags_for_resource}
\title{Retrieves the list of tags for an AWS IoT SiteWise resource}
\usage{
iotsitewise_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the resource.}
}
\description{
Retrieves the list of tags for an AWS IoT SiteWise resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
