% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_put_default_encryption_configuration}
\alias{iotsitewise_put_default_encryption_configuration}
\title{Sets the default encryption configuration for the AWS account}
\usage{
iotsitewise_put_default_encryption_configuration(encryptionType,
  kmsKeyId)
}
\arguments{
\item{encryptionType}{[required] The type of encryption used for the encryption configuration.}

\item{kmsKeyId}{The Key ID of the customer managed customer master key (CMK) used for
AWS KMS encryption. This is required if you use \code{KMS_BASED_ENCRYPTION}.}
}
\description{
Sets the default encryption configuration for the AWS account. For more
information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/key-management.html}{Key management}
in the \emph{AWS IoT SiteWise User Guide}.
}
\section{Request syntax}{
\preformatted{svc$put_default_encryption_configuration(
  encryptionType = "SITEWISE_DEFAULT_ENCRYPTION"|"KMS_BASED_ENCRYPTION",
  kmsKeyId = "string"
)
}
}

\keyword{internal}
