% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_put_logging_options}
\alias{iotsitewise_put_logging_options}
\title{Sets logging options for AWS IoT SiteWise}
\usage{
iotsitewise_put_logging_options(loggingOptions)
}
\arguments{
\item{loggingOptions}{[required] The logging options to set.}
}
\description{
Sets logging options for AWS IoT SiteWise.
}
\section{Request syntax}{
\preformatted{svc$put_logging_options(
  loggingOptions = list(
    level = "ERROR"|"INFO"|"OFF"
  )
)
}
}

\keyword{internal}
