% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_untag_resource}
\alias{iotsitewise_untag_resource}
\title{Removes a tag from an AWS IoT SiteWise resource}
\usage{
iotsitewise_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of the resource to untag.}

\item{tagKeys}{[required] A list of keys for tags to remove from the resource.}
}
\description{
Removes a tag from an AWS IoT SiteWise resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
