% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_update_asset}
\alias{iotsitewise_update_asset}
\title{Updates an asset's name}
\usage{
iotsitewise_update_asset(assetId, assetName, clientToken)
}
\arguments{
\item{assetId}{[required] The ID of the asset to update.}

\item{assetName}{[required] A unique, friendly name for the asset.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don't reuse this client token if a new
idempotent request is required.}
}
\description{
Updates an asset's name. For more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/update-assets-and-models.html}{Updating assets and models}
in the \emph{AWS IoT SiteWise User Guide}.
}
\section{Request syntax}{
\preformatted{svc$update_asset(
  assetId = "string",
  assetName = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
