% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_update_asset_property}
\alias{iotsitewise_update_asset_property}
\title{Updates an asset property's alias and notification state}
\usage{
iotsitewise_update_asset_property(assetId, propertyId, propertyAlias,
  propertyNotificationState, clientToken)
}
\arguments{
\item{assetId}{[required] The ID of the asset to be updated.}

\item{propertyId}{[required] The ID of the asset property to be updated.}

\item{propertyAlias}{The property alias that identifies the property, such as an OPC-UA
server data stream path (for example,
\verb{/company/windfarm/3/turbine/7/temperature}). For more information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/connect-data-streams.html}{Mapping industrial data streams to asset properties}
in the \emph{AWS IoT SiteWise User Guide}.

If you omit this parameter, the alias is removed from the property.}

\item{propertyNotificationState}{The MQTT notification state (enabled or disabled) for this asset
property. When the notification state is enabled, AWS IoT SiteWise
publishes property value updates to a unique MQTT topic. For more
information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/interact-with-other-services.html}{Interacting with other services}
in the \emph{AWS IoT SiteWise User Guide}.

If you omit this parameter, the notification state is set to \code{DISABLED}.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don't reuse this client token if a new
idempotent request is required.}
}
\description{
Updates an asset property's alias and notification state.

This operation overwrites the property's existing alias and notification
state. To keep your existing property's alias or notification state, you
must include the existing values in the UpdateAssetProperty request. For
more information, see
\code{\link[=iotsitewise_describe_asset_property]{describe_asset_property}}.
}
\section{Request syntax}{
\preformatted{svc$update_asset_property(
  assetId = "string",
  propertyId = "string",
  propertyAlias = "string",
  propertyNotificationState = "ENABLED"|"DISABLED",
  clientToken = "string"
)
}
}

\keyword{internal}
