% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_update_dashboard}
\alias{iotsitewise_update_dashboard}
\title{Updates an AWS IoT SiteWise Monitor dashboard}
\usage{
iotsitewise_update_dashboard(dashboardId, dashboardName,
  dashboardDescription, dashboardDefinition, clientToken)
}
\arguments{
\item{dashboardId}{[required] The ID of the dashboard to update.}

\item{dashboardName}{[required] A new friendly name for the dashboard.}

\item{dashboardDescription}{A new description for the dashboard.}

\item{dashboardDefinition}{[required] The new dashboard definition, as specified in a JSON literal. For
detailed information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/create-dashboards-using-aws-cli.html}{Creating dashboards (CLI)}
in the \emph{AWS IoT SiteWise User Guide}.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don't reuse this client token if a new
idempotent request is required.}
}
\description{
Updates an AWS IoT SiteWise Monitor dashboard.
}
\section{Request syntax}{
\preformatted{svc$update_dashboard(
  dashboardId = "string",
  dashboardName = "string",
  dashboardDescription = "string",
  dashboardDefinition = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
