% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_update_gateway}
\alias{iotsitewise_update_gateway}
\title{Updates a gateway's name}
\usage{
iotsitewise_update_gateway(gatewayId, gatewayName)
}
\arguments{
\item{gatewayId}{[required] The ID of the gateway to update.}

\item{gatewayName}{[required] A unique, friendly name for the gateway.}
}
\description{
Updates a gateway's name.
}
\section{Request syntax}{
\preformatted{svc$update_gateway(
  gatewayId = "string",
  gatewayName = "string"
)
}
}

\keyword{internal}
