% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_update_gateway_capability_configuration}
\alias{iotsitewise_update_gateway_capability_configuration}
\title{Updates a gateway capability configuration or defines a new capability
configuration}
\usage{
iotsitewise_update_gateway_capability_configuration(gatewayId,
  capabilityNamespace, capabilityConfiguration)
}
\arguments{
\item{gatewayId}{[required] The ID of the gateway to be updated.}

\item{capabilityNamespace}{[required] The namespace of the gateway capability configuration to be updated. For
example, if you configure OPC-UA sources from the AWS IoT SiteWise
console, your OPC-UA capability configuration has the namespace
\code{iotsitewise:opcuacollector:version}, where \code{version} is a number such
as \code{1}.}

\item{capabilityConfiguration}{[required] The JSON document that defines the configuration for the gateway
capability. For more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/configure-sources.html#configure-source-cli}{Configuring data sources (CLI)}
in the \emph{AWS IoT SiteWise User Guide}.}
}
\description{
Updates a gateway capability configuration or defines a new capability
configuration. Each gateway capability defines data sources for a
gateway. A capability configuration can contain multiple data source
configurations. If you define OPC-UA sources for a gateway in the AWS
IoT SiteWise console, all of your OPC-UA sources are stored in one
capability configuration. To list all capability configurations for a
gateway, use \code{\link[=iotsitewise_describe_gateway]{describe_gateway}}.
}
\section{Request syntax}{
\preformatted{svc$update_gateway_capability_configuration(
  gatewayId = "string",
  capabilityNamespace = "string",
  capabilityConfiguration = "string"
)
}
}

\keyword{internal}
