% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_update_portal}
\alias{iotsitewise_update_portal}
\title{Updates an AWS IoT SiteWise Monitor portal}
\usage{
iotsitewise_update_portal(portalId, portalName, portalDescription,
  portalContactEmail, portalLogoImage, roleArn, clientToken)
}
\arguments{
\item{portalId}{[required] The ID of the portal to update.}

\item{portalName}{[required] A new friendly name for the portal.}

\item{portalDescription}{A new description for the portal.}

\item{portalContactEmail}{[required] The AWS administrator's contact email address.}

\item{portalLogoImage}{}

\item{roleArn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of a service role that allows the portal's users to access your AWS IoT
SiteWise resources on your behalf. For more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/monitor-service-role.html}{Using service roles for AWS IoT SiteWise Monitor}
in the \emph{AWS IoT SiteWise User Guide}.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don't reuse this client token if a new
idempotent request is required.}
}
\description{
Updates an AWS IoT SiteWise Monitor portal.
}
\section{Request syntax}{
\preformatted{svc$update_portal(
  portalId = "string",
  portalName = "string",
  portalDescription = "string",
  portalContactEmail = "string",
  portalLogoImage = list(
    id = "string",
    file = list(
      data = raw,
      type = "PNG"
    )
  ),
  roleArn = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
