% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_update_project}
\alias{iotsitewise_update_project}
\title{Updates an AWS IoT SiteWise Monitor project}
\usage{
iotsitewise_update_project(projectId, projectName, projectDescription,
  clientToken)
}
\arguments{
\item{projectId}{[required] The ID of the project to update.}

\item{projectName}{[required] A new friendly name for the project.}

\item{projectDescription}{A new description for the project.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don\'t reuse this client token if a new
idempotent request is required.}
}
\description{
Updates an AWS IoT SiteWise Monitor project.
}
\section{Request syntax}{
\preformatted{svc$update_project(
  projectId = "string",
  projectName = "string",
  projectDescription = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
