% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_associate_entity_to_thing}
\alias{iotthingsgraph_associate_entity_to_thing}
\title{Associates a device with a concrete thing that is in the user's registry}
\usage{
iotthingsgraph_associate_entity_to_thing(thingName, entityId,
  namespaceVersion)
}
\arguments{
\item{thingName}{[required] The name of the thing to which the entity is to be associated.}

\item{entityId}{[required] The ID of the device to be associated with the thing.

The ID should be in the following format.

\verb{urn:tdm:REGION/ACCOUNT ID/default:device:DEVICENAME}}

\item{namespaceVersion}{The version of the user\'s namespace. Defaults to the latest version of
the user\'s namespace.}
}
\description{
Associates a device with a concrete thing that is in the user\'s
registry.
}
\details{
A thing can be associated with only one device at a time. If you
associate a thing with a new device id, its previous association will be
removed.
}
\section{Request syntax}{
\preformatted{svc$associate_entity_to_thing(
  thingName = "string",
  entityId = "string",
  namespaceVersion = 123
)
}
}

\keyword{internal}
