% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_create_system_instance}
\alias{iotthingsgraph_create_system_instance}
\title{Creates a system instance}
\usage{
iotthingsgraph_create_system_instance(tags, definition, target,
  greengrassGroupName, s3BucketName, metricsConfiguration,
  flowActionsRoleArn)
}
\arguments{
\item{tags}{Metadata, consisting of key-value pairs, that can be used to categorize
your system instances.}

\item{definition}{[required]}

\item{target}{[required] The target type of the deployment. Valid values are \code{GREENGRASS} and
\code{CLOUD}.}

\item{greengrassGroupName}{The name of the Greengrass group where the system instance will be
deployed. This value is required if the value of the \code{target} parameter
is \code{GREENGRASS}.}

\item{s3BucketName}{The name of the Amazon Simple Storage Service bucket that will be used
to store and deploy the system instance\'s resource file. This value is
required if the value of the \code{target} parameter is \code{GREENGRASS}.}

\item{metricsConfiguration}{}

\item{flowActionsRoleArn}{The ARN of the IAM role that AWS IoT Things Graph will assume when it
executes the flow. This role must have read and write access to AWS
Lambda and AWS IoT and any other AWS services that the flow uses when it
executes. This value is required if the value of the \code{target} parameter
is \code{CLOUD}.}
}
\description{
Creates a system instance.
}
\details{
This action validates the system instance, prepares the
deployment-related resources. For Greengrass deployments, it updates the
Greengrass group that is specified by the \code{greengrassGroupName}
parameter. It also adds a file to the S3 bucket specified by the
\code{s3BucketName} parameter. You need to call \code{DeploySystemInstance} after
running this action.

For Greengrass deployments, since this action modifies and adds
resources to a Greengrass group and an S3 bucket on the caller\'s
behalf, the calling identity must have write permissions to both the
specified Greengrass group and S3 bucket. Otherwise, the call will fail
with an authorization error.

For cloud deployments, this action requires a \code{flowActionsRoleArn}
value. This is an IAM role that has permissions to access AWS services,
such as AWS Lambda and AWS IoT, that the flow uses when it executes.

If the definition document doesn\'t specify a version of the user\'s
namespace, the latest version will be used by default.
}
\section{Request syntax}{
\preformatted{svc$create_system_instance(
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  ),
  definition = list(
    language = "GRAPHQL",
    text = "string"
  ),
  target = "GREENGRASS"|"CLOUD",
  greengrassGroupName = "string",
  s3BucketName = "string",
  metricsConfiguration = list(
    cloudMetricEnabled = TRUE|FALSE,
    metricRuleRoleArn = "string"
  ),
  flowActionsRoleArn = "string"
)
}
}

\keyword{internal}
