% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_create_system_template}
\alias{iotthingsgraph_create_system_template}
\title{Creates a system}
\usage{
iotthingsgraph_create_system_template(definition,
  compatibleNamespaceVersion)
}
\arguments{
\item{definition}{[required] The \code{DefinitionDocument} used to create the system.}

\item{compatibleNamespaceVersion}{The namespace version in which the system is to be created.

If no value is specified, the latest version is used by default.}
}
\description{
Creates a system. The system is validated against the entities in the
latest version of the user\'s namespace unless another namespace version
is specified in the request.
}
\section{Request syntax}{
\preformatted{svc$create_system_template(
  definition = list(
    language = "GRAPHQL",
    text = "string"
  ),
  compatibleNamespaceVersion = 123
)
}
}

\keyword{internal}
