% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_delete_flow_template}
\alias{iotthingsgraph_delete_flow_template}
\title{Deletes a workflow}
\usage{
iotthingsgraph_delete_flow_template(id)
}
\arguments{
\item{id}{[required] The ID of the workflow to be deleted.

The ID should be in the following format.

\verb{urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME}}
}
\description{
Deletes a workflow. Any new system or deployment that contains this
workflow will fail to update or deploy. Existing deployments that
contain the workflow will continue to run (since they use a snapshot of
the workflow taken at the time of deployment).
}
\section{Request syntax}{
\preformatted{svc$delete_flow_template(
  id = "string"
)
}
}

\keyword{internal}
