% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_delete_namespace}
\alias{iotthingsgraph_delete_namespace}
\title{Deletes the specified namespace}
\usage{
iotthingsgraph_delete_namespace()
}
\description{
Deletes the specified namespace. This action deletes all of the entities
in the namespace. Delete the systems and flows that use entities in the
namespace before performing this action.
}
\section{Request syntax}{
\preformatted{svc$delete_namespace()
}
}

\keyword{internal}
