% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_delete_system_template}
\alias{iotthingsgraph_delete_system_template}
\title{Deletes a system}
\usage{
iotthingsgraph_delete_system_template(id)
}
\arguments{
\item{id}{[required] The ID of the system to be deleted.

The ID should be in the following format.

\verb{urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME}}
}
\description{
Deletes a system. New deployments can't contain the system after its
deletion. Existing deployments that contain the system will continue to
work because they use a snapshot of the system that is taken when it is
deployed.
}
\section{Request syntax}{
\preformatted{svc$delete_system_template(
  id = "string"
)
}
}

\keyword{internal}
