% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_deprecate_flow_template}
\alias{iotthingsgraph_deprecate_flow_template}
\title{Deprecates the specified workflow}
\usage{
iotthingsgraph_deprecate_flow_template(id)
}
\arguments{
\item{id}{[required] The ID of the workflow to be deleted.

The ID should be in the following format.

\code{urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME}}
}
\description{
Deprecates the specified workflow. This action marks the workflow for
deletion. Deprecated flows can't be deployed, but existing deployments
will continue to run.
}
\section{Request syntax}{
\preformatted{svc$deprecate_flow_template(
  id = "string"
)
}
}

\keyword{internal}
