% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_deprecate_system_template}
\alias{iotthingsgraph_deprecate_system_template}
\title{Deprecates the specified system}
\usage{
iotthingsgraph_deprecate_system_template(id)
}
\arguments{
\item{id}{[required] The ID of the system to delete.

The ID should be in the following format.

\verb{urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME}}
}
\description{
Deprecates the specified system.
}
\section{Request syntax}{
\preformatted{svc$deprecate_system_template(
  id = "string"
)
}
}

\keyword{internal}
