% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_describe_namespace}
\alias{iotthingsgraph_describe_namespace}
\title{Gets the latest version of the user's namespace and the public version
that it is tracking}
\usage{
iotthingsgraph_describe_namespace(namespaceName)
}
\arguments{
\item{namespaceName}{The name of the user's namespace. Set this to \code{aws} to get the public
namespace.}
}
\description{
Gets the latest version of the user's namespace and the public version
that it is tracking.
}
\section{Request syntax}{
\preformatted{svc$describe_namespace(
  namespaceName = "string"
)
}
}

\keyword{internal}
