% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_dissociate_entity_from_thing}
\alias{iotthingsgraph_dissociate_entity_from_thing}
\title{Dissociates a device entity from a concrete thing}
\usage{
iotthingsgraph_dissociate_entity_from_thing(thingName, entityType)
}
\arguments{
\item{thingName}{[required] The name of the thing to disassociate.}

\item{entityType}{[required] The entity type from which to disassociate the thing.}
}
\description{
Dissociates a device entity from a concrete thing. The action takes only
the type of the entity that you need to dissociate because only one
entity of a particular type can be associated with a thing.
}
\section{Request syntax}{
\preformatted{svc$dissociate_entity_from_thing(
  thingName = "string",
  entityType = "DEVICE"|"SERVICE"|"DEVICE_MODEL"|"CAPABILITY"|"STATE"|"ACTION"|"EVENT"|"PROPERTY"|"MAPPING"|"ENUM"
)
}
}

\keyword{internal}
