% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_get_entities}
\alias{iotthingsgraph_get_entities}
\title{Gets definitions of the specified entities}
\usage{
iotthingsgraph_get_entities(ids, namespaceVersion)
}
\arguments{
\item{ids}{[required] An array of entity IDs.

The IDs should be in the following format.

\verb{urn:tdm:REGION/ACCOUNT ID/default:device:DEVICENAME}}

\item{namespaceVersion}{The version of the user's namespace. Defaults to the latest version of
the user's namespace.}
}
\description{
Gets definitions of the specified entities. Uses the latest version of
the user's namespace by default. This API returns the following TDM
entities.
\itemize{
\item Properties
\item States
\item Events
\item Actions
\item Capabilities
\item Mappings
\item Devices
\item Device Models
\item Services
}

This action doesn't return definitions for systems, flows, and
deployments.
}
\section{Request syntax}{
\preformatted{svc$get_entities(
  ids = list(
    "string"
  ),
  namespaceVersion = 123
)
}
}

\keyword{internal}
