% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_get_flow_template}
\alias{iotthingsgraph_get_flow_template}
\title{Gets the latest version of the DefinitionDocument and
FlowTemplateSummary for the specified workflow}
\usage{
iotthingsgraph_get_flow_template(id, revisionNumber)
}
\arguments{
\item{id}{[required] The ID of the workflow.

The ID should be in the following format.

\code{urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME}}

\item{revisionNumber}{The number of the workflow revision to retrieve.}
}
\description{
Gets the latest version of the \code{DefinitionDocument} and
\code{FlowTemplateSummary} for the specified workflow.
}
\section{Request syntax}{
\preformatted{svc$get_flow_template(
  id = "string",
  revisionNumber = 123
)
}
}

\keyword{internal}
