% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_get_flow_template_revisions}
\alias{iotthingsgraph_get_flow_template_revisions}
\title{Gets revisions of the specified workflow}
\usage{
iotthingsgraph_get_flow_template_revisions(id, nextToken, maxResults)
}
\arguments{
\item{id}{[required] The ID of the workflow.

The ID should be in the following format.

\verb{urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME}}

\item{nextToken}{The string that specifies the next page of results. Use this when
you\'re paginating results.}

\item{maxResults}{The maximum number of results to return in the response.}
}
\description{
Gets revisions of the specified workflow. Only the last 100 revisions
are stored. If the workflow has been deprecated, this action will return
revisions that occurred before the deprecation. This action won\'t work
for workflows that have been deleted.
}
\section{Request syntax}{
\preformatted{svc$get_flow_template_revisions(
  id = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
