% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_get_namespace_deletion_status}
\alias{iotthingsgraph_get_namespace_deletion_status}
\title{Gets the status of a namespace deletion task}
\usage{
iotthingsgraph_get_namespace_deletion_status()
}
\description{
Gets the status of a namespace deletion task.
}
\section{Request syntax}{
\preformatted{svc$get_namespace_deletion_status()
}
}

\keyword{internal}
