% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_get_system_instance}
\alias{iotthingsgraph_get_system_instance}
\title{Gets a system instance}
\usage{
iotthingsgraph_get_system_instance(id)
}
\arguments{
\item{id}{[required] The ID of the system deployment instance. This value is returned by
\code{CreateSystemInstance}.

The ID should be in the following format.

\code{urn:tdm:REGION/ACCOUNT ID/default:deployment:DEPLOYMENTNAME}}
}
\description{
Gets a system instance.
}
\section{Request syntax}{
\preformatted{svc$get_system_instance(
  id = "string"
)
}
}

\keyword{internal}
