% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_get_system_template}
\alias{iotthingsgraph_get_system_template}
\title{Gets a system}
\usage{
iotthingsgraph_get_system_template(id, revisionNumber)
}
\arguments{
\item{id}{[required] The ID of the system to get. This ID must be in the user\'s namespace.

The ID should be in the following format.

\verb{urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME}}

\item{revisionNumber}{The number that specifies the revision of the system to get.}
}
\description{
Gets a system.
}
\section{Request syntax}{
\preformatted{svc$get_system_template(
  id = "string",
  revisionNumber = 123
)
}
}

\keyword{internal}
