% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_get_upload_status}
\alias{iotthingsgraph_get_upload_status}
\title{Gets the status of the specified upload}
\usage{
iotthingsgraph_get_upload_status(uploadId)
}
\arguments{
\item{uploadId}{[required] The ID of the upload. This value is returned by the
\code{UploadEntityDefinitions} action.}
}
\description{
Gets the status of the specified upload.
}
\section{Request syntax}{
\preformatted{svc$get_upload_status(
  uploadId = "string"
)
}
}

\keyword{internal}
