% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_list_flow_execution_messages}
\alias{iotthingsgraph_list_flow_execution_messages}
\title{Returns a list of objects that contain information about events in a
flow execution}
\usage{
iotthingsgraph_list_flow_execution_messages(flowExecutionId, nextToken,
  maxResults)
}
\arguments{
\item{flowExecutionId}{[required] The ID of the flow execution.}

\item{nextToken}{The string that specifies the next page of results. Use this when
you\'re paginating results.}

\item{maxResults}{The maximum number of results to return in the response.}
}
\description{
Returns a list of objects that contain information about events in a
flow execution.
}
\section{Request syntax}{
\preformatted{svc$list_flow_execution_messages(
  flowExecutionId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
