% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_list_tags_for_resource}
\alias{iotthingsgraph_list_tags_for_resource}
\title{Lists all tags on an AWS IoT Things Graph resource}
\usage{
iotthingsgraph_list_tags_for_resource(maxResults, resourceArn,
  nextToken)
}
\arguments{
\item{maxResults}{The maximum number of tags to return.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource whose tags are to be
returned.}

\item{nextToken}{The token that specifies the next page of results to return.}
}
\description{
Lists all tags on an AWS IoT Things Graph resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  maxResults = 123,
  resourceArn = "string",
  nextToken = "string"
)
}
}

\keyword{internal}
