% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_search_entities}
\alias{iotthingsgraph_search_entities}
\title{Searches for entities of the specified type}
\usage{
iotthingsgraph_search_entities(entityTypes, filters, nextToken,
  maxResults, namespaceVersion)
}
\arguments{
\item{entityTypes}{[required] The entity types for which to search.}

\item{filters}{Optional filter to apply to the search. Valid filters are \code{NAME}
\code{NAMESPACE}, \code{SEMANTIC_TYPE_PATH} and \code{REFERENCED_ENTITY_ID}.
\code{REFERENCED_ENTITY_ID} filters on entities that are used by the entity
in the result set. For example, you can filter on the ID of a property
that is used in a state.

Multiple filters function as OR criteria in the query. Multiple values
passed inside the filter function as AND criteria.}

\item{nextToken}{The string that specifies the next page of results. Use this when
you're paginating results.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{namespaceVersion}{The version of the user's namespace. Defaults to the latest version of
the user's namespace.}
}
\description{
Searches for entities of the specified type. You can search for entities
in your namespace and the public namespace that you're tracking.
}
\section{Request syntax}{
\preformatted{svc$search_entities(
  entityTypes = list(
    "DEVICE"|"SERVICE"|"DEVICE_MODEL"|"CAPABILITY"|"STATE"|"ACTION"|"EVENT"|"PROPERTY"|"MAPPING"|"ENUM"
  ),
  filters = list(
    list(
      name = "NAME"|"NAMESPACE"|"SEMANTIC_TYPE_PATH"|"REFERENCED_ENTITY_ID",
      value = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123,
  namespaceVersion = 123
)
}
}

\keyword{internal}
