% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_search_flow_executions}
\alias{iotthingsgraph_search_flow_executions}
\title{Searches for AWS IoT Things Graph workflow execution instances}
\usage{
iotthingsgraph_search_flow_executions(systemInstanceId, flowExecutionId,
  startTime, endTime, nextToken, maxResults)
}
\arguments{
\item{systemInstanceId}{[required] The ID of the system instance that contains the flow.}

\item{flowExecutionId}{The ID of a flow execution.}

\item{startTime}{The date and time of the earliest flow execution to return.}

\item{endTime}{The date and time of the latest flow execution to return.}

\item{nextToken}{The string that specifies the next page of results. Use this when you're
paginating results.}

\item{maxResults}{The maximum number of results to return in the response.}
}
\description{
Searches for AWS IoT Things Graph workflow execution instances.
}
\section{Request syntax}{
\preformatted{svc$search_flow_executions(
  systemInstanceId = "string",
  flowExecutionId = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
