% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_search_flow_templates}
\alias{iotthingsgraph_search_flow_templates}
\title{Searches for summary information about workflows}
\usage{
iotthingsgraph_search_flow_templates(filters, nextToken, maxResults)
}
\arguments{
\item{filters}{An array of objects that limit the result set. The only valid filter is
\code{DEVICE_MODEL_ID}.}

\item{nextToken}{The string that specifies the next page of results. Use this when
you're paginating results.}

\item{maxResults}{The maximum number of results to return in the response.}
}
\description{
Searches for summary information about workflows.
}
\section{Request syntax}{
\preformatted{svc$search_flow_templates(
  filters = list(
    list(
      name = "DEVICE_MODEL_ID",
      value = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
