% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_search_system_instances}
\alias{iotthingsgraph_search_system_instances}
\title{Searches for system instances in the user's account}
\usage{
iotthingsgraph_search_system_instances(filters, nextToken, maxResults)
}
\arguments{
\item{filters}{Optional filter to apply to the search. Valid filters are
\code{SYSTEM_TEMPLATE_ID}, \code{STATUS}, and \code{GREENGRASS_GROUP_NAME}.

Multiple filters function as OR criteria in the query. Multiple values
passed inside the filter function as AND criteria.}

\item{nextToken}{The string that specifies the next page of results. Use this when
you're paginating results.}

\item{maxResults}{The maximum number of results to return in the response.}
}
\description{
Searches for system instances in the user's account.
}
\section{Request syntax}{
\preformatted{svc$search_system_instances(
  filters = list(
    list(
      name = "SYSTEM_TEMPLATE_ID"|"STATUS"|"GREENGRASS_GROUP_NAME",
      value = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
