% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_search_system_templates}
\alias{iotthingsgraph_search_system_templates}
\title{Searches for summary information about systems in the user's account}
\usage{
iotthingsgraph_search_system_templates(filters, nextToken, maxResults)
}
\arguments{
\item{filters}{An array of filters that limit the result set. The only valid filter is
\code{FLOW_TEMPLATE_ID}.}

\item{nextToken}{The string that specifies the next page of results. Use this when
you're paginating results.}

\item{maxResults}{The maximum number of results to return in the response.}
}
\description{
Searches for summary information about systems in the user's account.
You can filter by the ID of a workflow to return only systems that use
the specified workflow.
}
\section{Request syntax}{
\preformatted{svc$search_system_templates(
  filters = list(
    list(
      name = "FLOW_TEMPLATE_ID",
      value = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
