% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_update_flow_template}
\alias{iotthingsgraph_update_flow_template}
\title{Updates the specified workflow}
\usage{
iotthingsgraph_update_flow_template(id, definition,
  compatibleNamespaceVersion)
}
\arguments{
\item{id}{[required] The ID of the workflow to be updated.

The ID should be in the following format.

\verb{urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME}}

\item{definition}{[required] The \code{DefinitionDocument} that contains the updated workflow definition.}

\item{compatibleNamespaceVersion}{The version of the user\'s namespace.

If no value is specified, the latest version is used by default. Use the
\code{GetFlowTemplateRevisions} if you want to find earlier revisions of the
flow to update.}
}
\description{
Updates the specified workflow. All deployed systems and system
instances that use the workflow will see the changes in the flow when it
is redeployed. If you don\'t want this behavior, copy the workflow
(creating a new workflow with a different ID), and update the copy. The
workflow can contain only entities in the specified namespace.
}
\section{Request syntax}{
\preformatted{svc$update_flow_template(
  id = "string",
  definition = list(
    language = "GRAPHQL",
    text = "string"
  ),
  compatibleNamespaceVersion = 123
)
}
}

\keyword{internal}
