% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_update_system_template}
\alias{iotthingsgraph_update_system_template}
\title{Updates the specified system}
\usage{
iotthingsgraph_update_system_template(id, definition,
  compatibleNamespaceVersion)
}
\arguments{
\item{id}{[required] The ID of the system to be updated.

The ID should be in the following format.

\code{urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME}}

\item{definition}{[required] The \code{DefinitionDocument} that contains the updated system definition.}

\item{compatibleNamespaceVersion}{The version of the user's namespace. Defaults to the latest version of
the user's namespace.

If no value is specified, the latest version is used by default.}
}
\description{
Updates the specified system. You don't need to run this action after
updating a workflow. Any deployment that uses the system will see the
changes in the system when it is redeployed.
}
\section{Request syntax}{
\preformatted{svc$update_system_template(
  id = "string",
  definition = list(
    language = "GRAPHQL",
    text = "string"
  ),
  compatibleNamespaceVersion = 123
)
}
}

\keyword{internal}
