% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotthingsgraph_operations.R
\name{iotthingsgraph_upload_entity_definitions}
\alias{iotthingsgraph_upload_entity_definitions}
\title{Asynchronously uploads one or more entity definitions to the user's
namespace}
\usage{
iotthingsgraph_upload_entity_definitions(document,
  syncWithPublicNamespace, deprecateExistingEntities)
}
\arguments{
\item{document}{The \code{DefinitionDocument} that defines the updated entities.}

\item{syncWithPublicNamespace}{A Boolean that specifies whether to synchronize with the latest version
of the public namespace. If set to \code{true}, the upload will create a new
namespace version.}

\item{deprecateExistingEntities}{A Boolean that specifies whether to deprecate all entities in the latest
version before uploading the new \code{DefinitionDocument}. If set to \code{true},
the upload will create a new namespace version.}
}
\description{
Asynchronously uploads one or more entity definitions to the user's
namespace. The \code{document} parameter is required if
\code{syncWithPublicNamespace} and \code{deleteExistingEntites} are false. If the
\code{syncWithPublicNamespace} parameter is set to \code{true}, the user's
namespace will synchronize with the latest version of the public
namespace. If \code{deprecateExistingEntities} is set to true, all entities
in the latest version will be deleted before the new
\code{DefinitionDocument} is uploaded.
}
\details{
When a user uploads entity definitions for the first time, the service
creates a new namespace for the user. The new namespace tracks the
public namespace. Currently users can have only one namespace. The
namespace version increments whenever a user uploads entity definitions
that are backwards-incompatible and whenever a user sets the
\code{syncWithPublicNamespace} parameter or the \code{deprecateExistingEntities}
parameter to \code{true}.

The IDs for all of the entities should be in URN format. Each entity
must be in the user's namespace. Users can't create entities in the
public namespace, but entity definitions can refer to entities in the
public namespace.

Valid entities are \code{Device}, \code{DeviceModel}, \code{Service}, \code{Capability},
\code{State}, \code{Action}, \code{Event}, \code{Property}, \code{Mapping}, \code{Enum}.
}
\section{Request syntax}{
\preformatted{svc$upload_entity_definitions(
  document = list(
    language = "GRAPHQL",
    text = "string"
  ),
  syncWithPublicNamespace = TRUE|FALSE,
  deprecateExistingEntities = TRUE|FALSE
)
}
}

\keyword{internal}
