% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_associate_aws_account_with_partner_account}
\alias{iotwireless_associate_aws_account_with_partner_account}
\title{Associates a partner account with your AWS account}
\usage{
iotwireless_associate_aws_account_with_partner_account(Sidewalk,
  ClientRequestToken)
}
\arguments{
\item{Sidewalk}{[required] The Sidewalk account credentials.}

\item{ClientRequestToken}{Each resource must have a unique client request token. If you try to
create a new resource with the same token as a resource that already
exists, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}
}
\description{
Associates a partner account with your AWS account.
}
\section{Request syntax}{
\preformatted{svc$associate_aws_account_with_partner_account(
  Sidewalk = list(
    AmazonId = "string",
    AppServerPrivateKey = "string"
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
