% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_associate_wireless_gateway_with_certificate}
\alias{iotwireless_associate_wireless_gateway_with_certificate}
\title{Associates a wireless gateway with a certificate}
\usage{
iotwireless_associate_wireless_gateway_with_certificate(Id,
  IotCertificateId)
}
\arguments{
\item{Id}{[required] The ID of the resource to update.}

\item{IotCertificateId}{[required] The ID of the certificate to associate with the wireless gateway.}
}
\description{
Associates a wireless gateway with a certificate.
}
\section{Request syntax}{
\preformatted{svc$associate_wireless_gateway_with_certificate(
  Id = "string",
  IotCertificateId = "string"
)
}
}

\keyword{internal}
