% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_associate_wireless_gateway_with_thing}
\alias{iotwireless_associate_wireless_gateway_with_thing}
\title{Associates a wireless gateway with a thing}
\usage{
iotwireless_associate_wireless_gateway_with_thing(Id, ThingArn)
}
\arguments{
\item{Id}{[required] The ID of the resource to update.}

\item{ThingArn}{[required] The ARN of the thing to associate with the wireless gateway.}
}
\description{
Associates a wireless gateway with a thing.
}
\section{Request syntax}{
\preformatted{svc$associate_wireless_gateway_with_thing(
  Id = "string",
  ThingArn = "string"
)
}
}

\keyword{internal}
