% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_create_destination}
\alias{iotwireless_create_destination}
\title{Creates a new destination that maps a device message to an AWS IoT rule}
\usage{
iotwireless_create_destination(Name, ExpressionType, Expression,
  Description, RoleArn, Tags, ClientRequestToken)
}
\arguments{
\item{Name}{[required] The name of the new resource.}

\item{ExpressionType}{[required] The type of value in \code{Expression}.}

\item{Expression}{[required] The rule name or topic rule to send messages to.}

\item{Description}{The description of the new resource.}

\item{RoleArn}{[required] The ARN of the IAM Role that authorizes the destination.}

\item{Tags}{The tags to attach to the new destination. Tags are metadata that can be
used to manage a resource.}

\item{ClientRequestToken}{Each resource must have a unique client request token. If you try to
create a new resource with the same token as a resource that already
exists, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}
}
\description{
Creates a new destination that maps a device message to an AWS IoT rule.
}
\section{Request syntax}{
\preformatted{svc$create_destination(
  Name = "string",
  ExpressionType = "RuleName",
  Expression = "string",
  Description = "string",
  RoleArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
