% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_create_device_profile}
\alias{iotwireless_create_device_profile}
\title{Creates a new device profile}
\usage{
iotwireless_create_device_profile(Name, LoRaWAN, Tags,
  ClientRequestToken)
}
\arguments{
\item{Name}{The name of the new resource.}

\item{LoRaWAN}{The device profile information to use to create the device profile.}

\item{Tags}{The tags to attach to the new device profile Tags are metadata that can
be used to manage a resource.}

\item{ClientRequestToken}{Each resource must have a unique client request token. If you try to
create a new resource with the same token as a resource that already
exists, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}
}
\description{
Creates a new device profile.
}
\section{Request syntax}{
\preformatted{svc$create_device_profile(
  Name = "string",
  LoRaWAN = list(
    SupportsClassB = TRUE|FALSE,
    ClassBTimeout = 123,
    PingSlotPeriod = 123,
    PingSlotDr = 123,
    PingSlotFreq = 123,
    SupportsClassC = TRUE|FALSE,
    ClassCTimeout = 123,
    MacVersion = "string",
    RegParamsRevision = "string",
    RxDelay1 = 123,
    RxDrOffset1 = 123,
    RxDataRate2 = 123,
    RxFreq2 = 123,
    FactoryPresetFreqsList = list(
      123
    ),
    MaxEirp = 123,
    MaxDutyCycle = 123,
    RfRegion = "string",
    SupportsJoin = TRUE|FALSE,
    Supports32BitFCnt = TRUE|FALSE
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
