% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_create_service_profile}
\alias{iotwireless_create_service_profile}
\title{Creates a new service profile}
\usage{
iotwireless_create_service_profile(Name, LoRaWAN, Tags,
  ClientRequestToken)
}
\arguments{
\item{Name}{The name of the new resource.}

\item{LoRaWAN}{The service profile information to use to create the service profile.}

\item{Tags}{The tags to attach to the new service profile. Tags are metadata that
can be used to manage a resource.}

\item{ClientRequestToken}{Each resource must have a unique client request token. If you try to
create a new resource with the same token as a resource that already
exists, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}
}
\description{
Creates a new service profile.
}
\section{Request syntax}{
\preformatted{svc$create_service_profile(
  Name = "string",
  LoRaWAN = list(
    AddGwMetadata = TRUE|FALSE
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
