% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_create_wireless_device}
\alias{iotwireless_create_wireless_device}
\title{Provisions a wireless device}
\usage{
iotwireless_create_wireless_device(Type, Name, Description,
  DestinationName, ClientRequestToken, LoRaWAN)
}
\arguments{
\item{Type}{[required] The wireless device type.}

\item{Name}{The name of the new resource.}

\item{Description}{The description of the new resource.}

\item{DestinationName}{[required] The name of the destination to assign to the new wireless device.}

\item{ClientRequestToken}{Each resource must have a unique client request token. If you try to
create a new resource with the same token as a resource that already
exists, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}

\item{LoRaWAN}{The device configuration information to use to create the wireless
device.}
}
\description{
Provisions a wireless device.
}
\section{Request syntax}{
\preformatted{svc$create_wireless_device(
  Type = "Sidewalk"|"LoRaWAN",
  Name = "string",
  Description = "string",
  DestinationName = "string",
  ClientRequestToken = "string",
  LoRaWAN = list(
    DevEui = "string",
    DeviceProfileId = "string",
    ServiceProfileId = "string",
    OtaaV1_1 = list(
      AppKey = "string",
      NwkKey = "string",
      JoinEui = "string"
    ),
    OtaaV1_0_x = list(
      AppKey = "string",
      AppEui = "string"
    ),
    AbpV1_1 = list(
      DevAddr = "string",
      SessionKeys = list(
        FNwkSIntKey = "string",
        SNwkSIntKey = "string",
        NwkSEncKey = "string",
        AppSKey = "string"
      )
    ),
    AbpV1_0_x = list(
      DevAddr = "string",
      SessionKeys = list(
        NwkSKey = "string",
        AppSKey = "string"
      )
    )
  )
)
}
}

\keyword{internal}
