% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_create_wireless_gateway}
\alias{iotwireless_create_wireless_gateway}
\title{Provisions a wireless gateway}
\usage{
iotwireless_create_wireless_gateway(Name, Description, LoRaWAN, Tags,
  ClientRequestToken)
}
\arguments{
\item{Name}{The name of the new resource.}

\item{Description}{The description of the new resource.}

\item{LoRaWAN}{[required] The gateway configuration information to use to create the wireless
gateway.}

\item{Tags}{The tags to attach to the new wireless gateway. Tags are metadata that
can be used to manage a resource.}

\item{ClientRequestToken}{Each resource must have a unique client request token. If you try to
create a new resource with the same token as a resource that already
exists, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}
}
\description{
Provisions a wireless gateway.
}
\section{Request syntax}{
\preformatted{svc$create_wireless_gateway(
  Name = "string",
  Description = "string",
  LoRaWAN = list(
    GatewayEui = "string",
    RfRegion = "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
