% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_create_wireless_gateway_task_definition}
\alias{iotwireless_create_wireless_gateway_task_definition}
\title{Creates a gateway task definition}
\usage{
iotwireless_create_wireless_gateway_task_definition(AutoCreateTasks,
  Name, Update, ClientRequestToken)
}
\arguments{
\item{AutoCreateTasks}{[required] Whether to automatically create tasks using this task definition for all
gateways with the specified current version. If \code{false}, the task must
me created by calling \code{CreateWirelessGatewayTask}.}

\item{Name}{The name of the new resource.}

\item{Update}{Information about the gateways to update.}

\item{ClientRequestToken}{Each resource must have a unique client request token. If you try to
create a new resource with the same token as a resource that already
exists, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}
}
\description{
Creates a gateway task definition.
}
\section{Request syntax}{
\preformatted{svc$create_wireless_gateway_task_definition(
  AutoCreateTasks = TRUE|FALSE,
  Name = "string",
  Update = list(
    UpdateDataSource = "string",
    UpdateDataRole = "string",
    LoRaWAN = list(
      UpdateSignature = "string",
      SigKeyCrc = 123,
      CurrentVersion = list(
        PackageVersion = "string",
        Model = "string",
        Station = "string"
      ),
      UpdateVersion = list(
        PackageVersion = "string",
        Model = "string",
        Station = "string"
      )
    )
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
