% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_delete_destination}
\alias{iotwireless_delete_destination}
\title{Deletes a destination}
\usage{
iotwireless_delete_destination(Name)
}
\arguments{
\item{Name}{[required] The name of the resource to delete.}
}
\description{
Deletes a destination.
}
\section{Request syntax}{
\preformatted{svc$delete_destination(
  Name = "string"
)
}
}

\keyword{internal}
