% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_delete_device_profile}
\alias{iotwireless_delete_device_profile}
\title{Deletes a device profile}
\usage{
iotwireless_delete_device_profile(Id)
}
\arguments{
\item{Id}{[required] The ID of the resource to delete.}
}
\description{
Deletes a device profile.
}
\section{Request syntax}{
\preformatted{svc$delete_device_profile(
  Id = "string"
)
}
}

\keyword{internal}
