% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_delete_wireless_gateway_task}
\alias{iotwireless_delete_wireless_gateway_task}
\title{Deletes a wireless gateway task}
\usage{
iotwireless_delete_wireless_gateway_task(Id)
}
\arguments{
\item{Id}{[required] The ID of the resource to delete.}
}
\description{
Deletes a wireless gateway task.
}
\section{Request syntax}{
\preformatted{svc$delete_wireless_gateway_task(
  Id = "string"
)
}
}

\keyword{internal}
