% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_disassociate_aws_account_from_partner_account}
\alias{iotwireless_disassociate_aws_account_from_partner_account}
\title{Disassociates your AWS account from a partner account}
\usage{
iotwireless_disassociate_aws_account_from_partner_account(
  PartnerAccountId, PartnerType)
}
\arguments{
\item{PartnerAccountId}{[required] The partner account ID to disassociate from the AWS account.}

\item{PartnerType}{[required] The partner type.}
}
\description{
Disassociates your AWS account from a partner account. If
\code{PartnerAccountId} and \code{PartnerType} are \code{null}, disassociates your AWS
account from all partner accounts.
}
\section{Request syntax}{
\preformatted{svc$disassociate_aws_account_from_partner_account(
  PartnerAccountId = "string",
  PartnerType = "Sidewalk"
)
}
}

\keyword{internal}
