% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_disassociate_wireless_gateway_from_thing}
\alias{iotwireless_disassociate_wireless_gateway_from_thing}
\title{Disassociates a wireless gateway from its currently associated thing}
\usage{
iotwireless_disassociate_wireless_gateway_from_thing(Id)
}
\arguments{
\item{Id}{[required] The ID of the resource to update.}
}
\description{
Disassociates a wireless gateway from its currently associated thing.
}
\section{Request syntax}{
\preformatted{svc$disassociate_wireless_gateway_from_thing(
  Id = "string"
)
}
}

\keyword{internal}
