% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_get_destination}
\alias{iotwireless_get_destination}
\title{Gets information about a destination}
\usage{
iotwireless_get_destination(Name)
}
\arguments{
\item{Name}{[required] The name of the resource to get.}
}
\description{
Gets information about a destination.
}
\section{Request syntax}{
\preformatted{svc$get_destination(
  Name = "string"
)
}
}

\keyword{internal}
