% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_get_partner_account}
\alias{iotwireless_get_partner_account}
\title{Gets information about a partner account}
\usage{
iotwireless_get_partner_account(PartnerAccountId, PartnerType)
}
\arguments{
\item{PartnerAccountId}{[required] The partner account ID to disassociate from the AWS account.}

\item{PartnerType}{[required] The partner type.}
}
\description{
Gets information about a partner account. If \code{PartnerAccountId} and
\code{PartnerType} are \code{null}, returns all partner accounts.
}
\section{Request syntax}{
\preformatted{svc$get_partner_account(
  PartnerAccountId = "string",
  PartnerType = "Sidewalk"
)
}
}

\keyword{internal}
