% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_get_service_endpoint}
\alias{iotwireless_get_service_endpoint}
\title{Gets the account-specific endpoint for Configuration and Update Server
(CUPS) protocol or LoRaWAN Network Server (LNS) connections}
\usage{
iotwireless_get_service_endpoint(ServiceType)
}
\arguments{
\item{ServiceType}{The service type for which to get endpoint information about. Can be
\code{CUPS} for the Configuration and Update Server endpoint, or \code{LNS} for
the LoRaWAN Network Server endpoint.}
}
\description{
Gets the account-specific endpoint for Configuration and Update Server
(CUPS) protocol or LoRaWAN Network Server (LNS) connections.
}
\section{Request syntax}{
\preformatted{svc$get_service_endpoint(
  ServiceType = "CUPS"|"LNS"
)
}
}

\keyword{internal}
