% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_get_service_profile}
\alias{iotwireless_get_service_profile}
\title{Gets information about a service profile}
\usage{
iotwireless_get_service_profile(Id)
}
\arguments{
\item{Id}{[required] The ID of the resource to get.}
}
\description{
Gets information about a service profile.
}
\section{Request syntax}{
\preformatted{svc$get_service_profile(
  Id = "string"
)
}
}

\keyword{internal}
