% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_get_wireless_gateway}
\alias{iotwireless_get_wireless_gateway}
\title{Gets information about a wireless gateway}
\usage{
iotwireless_get_wireless_gateway(Identifier, IdentifierType)
}
\arguments{
\item{Identifier}{[required] The identifier of the wireless gateway to get.}

\item{IdentifierType}{[required] The type of identifier used in \code{identifier}.}
}
\description{
Gets information about a wireless gateway.
}
\section{Request syntax}{
\preformatted{svc$get_wireless_gateway(
  Identifier = "string",
  IdentifierType = "GatewayEui"|"WirelessGatewayId"|"ThingName"
)
}
}

\keyword{internal}
