% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_get_wireless_gateway_certificate}
\alias{iotwireless_get_wireless_gateway_certificate}
\title{Gets the ID of the certificate that is currently associated with a
wireless gateway}
\usage{
iotwireless_get_wireless_gateway_certificate(Id)
}
\arguments{
\item{Id}{[required] The ID of the resource to get.}
}
\description{
Gets the ID of the certificate that is currently associated with a
wireless gateway.
}
\section{Request syntax}{
\preformatted{svc$get_wireless_gateway_certificate(
  Id = "string"
)
}
}

\keyword{internal}
