% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_get_wireless_gateway_firmware_information}
\alias{iotwireless_get_wireless_gateway_firmware_information}
\title{Gets the firmware version and other information about a wireless gateway}
\usage{
iotwireless_get_wireless_gateway_firmware_information(Id)
}
\arguments{
\item{Id}{[required] The ID of the resource to get.}
}
\description{
Gets the firmware version and other information about a wireless
gateway.
}
\section{Request syntax}{
\preformatted{svc$get_wireless_gateway_firmware_information(
  Id = "string"
)
}
}

\keyword{internal}
