% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_get_wireless_gateway_statistics}
\alias{iotwireless_get_wireless_gateway_statistics}
\title{Gets operating information about a wireless gateway}
\usage{
iotwireless_get_wireless_gateway_statistics(WirelessGatewayId)
}
\arguments{
\item{WirelessGatewayId}{[required] The ID of the wireless gateway for which to get the data.}
}
\description{
Gets operating information about a wireless gateway.
}
\section{Request syntax}{
\preformatted{svc$get_wireless_gateway_statistics(
  WirelessGatewayId = "string"
)
}
}

\keyword{internal}
