% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_get_wireless_gateway_task}
\alias{iotwireless_get_wireless_gateway_task}
\title{Gets information about a wireless gateway task}
\usage{
iotwireless_get_wireless_gateway_task(Id)
}
\arguments{
\item{Id}{[required] The ID of the resource to get.}
}
\description{
Gets information about a wireless gateway task.
}
\section{Request syntax}{
\preformatted{svc$get_wireless_gateway_task(
  Id = "string"
)
}
}

\keyword{internal}
