% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_list_destinations}
\alias{iotwireless_list_destinations}
\title{Lists the destinations registered to your AWS account}
\usage{
iotwireless_list_destinations(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in this operation.}

\item{NextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}
}
\description{
Lists the destinations registered to your AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_destinations(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
