% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_list_partner_accounts}
\alias{iotwireless_list_partner_accounts}
\title{Lists the partner accounts associated with your AWS account}
\usage{
iotwireless_list_partner_accounts(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{MaxResults}{The maximum number of results to return in this operation.}
}
\description{
Lists the partner accounts associated with your AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_partner_accounts(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
