% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_list_tags_for_resource}
\alias{iotwireless_list_tags_for_resource}
\title{Lists the tags (metadata) you have assigned to the resource}
\usage{
iotwireless_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource for which to list tags.}
}
\description{
Lists the tags (metadata) you have assigned to the resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
