% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_list_wireless_devices}
\alias{iotwireless_list_wireless_devices}
\title{Lists the wireless devices registered to your AWS account}
\usage{
iotwireless_list_wireless_devices(MaxResults, NextToken,
  DestinationName, DeviceProfileId, ServiceProfileId, WirelessDeviceType)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in this operation.}

\item{NextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{DestinationName}{A filter to list only the wireless devices that use this destination.}

\item{DeviceProfileId}{A filter to list only the wireless devices that use this device profile.}

\item{ServiceProfileId}{A filter to list only the wireless devices that use this service
profile.}

\item{WirelessDeviceType}{A filter to list only the wireless devices that use this wireless device
type.}
}
\description{
Lists the wireless devices registered to your AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_wireless_devices(
  MaxResults = 123,
  NextToken = "string",
  DestinationName = "string",
  DeviceProfileId = "string",
  ServiceProfileId = "string",
  WirelessDeviceType = "Sidewalk"|"LoRaWAN"
)
}
}

\keyword{internal}
