% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_list_wireless_gateway_task_definitions}
\alias{iotwireless_list_wireless_gateway_task_definitions}
\title{List the wireless gateway tasks definitions registered to your AWS
account}
\usage{
iotwireless_list_wireless_gateway_task_definitions(MaxResults,
  NextToken, TaskDefinitionType)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in this operation.}

\item{NextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{TaskDefinitionType}{A filter to list only the wireless gateway task definitions that use
this task definition type.}
}
\description{
List the wireless gateway tasks definitions registered to your AWS
account.
}
\section{Request syntax}{
\preformatted{svc$list_wireless_gateway_task_definitions(
  MaxResults = 123,
  NextToken = "string",
  TaskDefinitionType = "UPDATE"
)
}
}

\keyword{internal}
