% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_list_wireless_gateways}
\alias{iotwireless_list_wireless_gateways}
\title{Lists the wireless gateways registered to your AWS account}
\usage{
iotwireless_list_wireless_gateways(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{MaxResults}{The maximum number of results to return in this operation.}
}
\description{
Lists the wireless gateways registered to your AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_wireless_gateways(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
