% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_send_data_to_wireless_device}
\alias{iotwireless_send_data_to_wireless_device}
\title{Sends a decrypted application data frame to a device}
\usage{
iotwireless_send_data_to_wireless_device(Id, TransmitMode, PayloadData,
  WirelessMetadata)
}
\arguments{
\item{Id}{[required] The ID of the wireless device to receive the data.}

\item{TransmitMode}{[required] The transmit mode to use to send data to the wireless device. Can be:
\code{0} for UM (unacknowledge mode), \code{1} for AM (acknowledge mode), or \code{2}
for (TM) transparent mode.}

\item{PayloadData}{[required] The message payload to send.}

\item{WirelessMetadata}{Metadata about the message request.}
}
\description{
Sends a decrypted application data frame to a device.
}
\section{Request syntax}{
\preformatted{svc$send_data_to_wireless_device(
  Id = "string",
  TransmitMode = 123,
  PayloadData = "string",
  WirelessMetadata = list(
    LoRaWAN = list(
      FPort = 123
    ),
    Sidewalk = list(
      Seq = 123
    )
  )
)
}
}

\keyword{internal}
