% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_tag_resource}
\alias{iotwireless_tag_resource}
\title{Adds a tag to a resource}
\usage{
iotwireless_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource to add tags to.}

\item{Tags}{[required] Adds to or modifies the tags of the given resource. Tags are metadata
that can be used to manage a resource.}
}
\description{
Adds a tag to a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
