% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_test_wireless_device}
\alias{iotwireless_test_wireless_device}
\title{Simulates a provisioned device by sending an uplink data payload of
Hello}
\usage{
iotwireless_test_wireless_device(Id)
}
\arguments{
\item{Id}{[required] The ID of the wireless device to test.}
}
\description{
Simulates a provisioned device by sending an uplink data payload of
\code{Hello}.
}
\section{Request syntax}{
\preformatted{svc$test_wireless_device(
  Id = "string"
)
}
}

\keyword{internal}
