% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_untag_resource}
\alias{iotwireless_untag_resource}
\title{Removes one or more tags from a resource}
\usage{
iotwireless_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource to remove tags from.}

\item{TagKeys}{[required] A list of the keys of the tags to remove from the resource.}
}
\description{
Removes one or more tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
