% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_update_destination}
\alias{iotwireless_update_destination}
\title{Updates properties of a destination}
\usage{
iotwireless_update_destination(Name, ExpressionType, Expression,
  Description, RoleArn)
}
\arguments{
\item{Name}{[required] The new name of the resource.}

\item{ExpressionType}{The type of value in \code{Expression}.}

\item{Expression}{The new rule name or topic rule to send messages to.}

\item{Description}{A new description of the resource.}

\item{RoleArn}{The ARN of the IAM Role that authorizes the destination.}
}
\description{
Updates properties of a destination.
}
\section{Request syntax}{
\preformatted{svc$update_destination(
  Name = "string",
  ExpressionType = "RuleName",
  Expression = "string",
  Description = "string",
  RoleArn = "string"
)
}
}

\keyword{internal}
