% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_update_partner_account}
\alias{iotwireless_update_partner_account}
\title{Updates properties of a partner account}
\usage{
iotwireless_update_partner_account(Sidewalk, PartnerAccountId,
  PartnerType)
}
\arguments{
\item{Sidewalk}{[required] The Sidewalk account credentials.}

\item{PartnerAccountId}{[required] The ID of the partner account to update.}

\item{PartnerType}{[required] The partner type.}
}
\description{
Updates properties of a partner account.
}
\section{Request syntax}{
\preformatted{svc$update_partner_account(
  Sidewalk = list(
    AppServerPrivateKey = "string"
  ),
  PartnerAccountId = "string",
  PartnerType = "Sidewalk"
)
}
}

\keyword{internal}
