% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_update_wireless_device}
\alias{iotwireless_update_wireless_device}
\title{Updates properties of a wireless device}
\usage{
iotwireless_update_wireless_device(Id, DestinationName, Name,
  Description, LoRaWAN)
}
\arguments{
\item{Id}{[required] The ID of the resource to update.}

\item{DestinationName}{The name of the new destination for the device.}

\item{Name}{The new name of the resource.}

\item{Description}{A new description of the resource.}

\item{LoRaWAN}{The updated wireless device's configuration.}
}
\description{
Updates properties of a wireless device.
}
\section{Request syntax}{
\preformatted{svc$update_wireless_device(
  Id = "string",
  DestinationName = "string",
  Name = "string",
  Description = "string",
  LoRaWAN = list(
    DeviceProfileId = "string",
    ServiceProfileId = "string"
  )
)
}
}

\keyword{internal}
