% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotwireless_operations.R
\name{iotwireless_update_wireless_gateway}
\alias{iotwireless_update_wireless_gateway}
\title{Updates properties of a wireless gateway}
\usage{
iotwireless_update_wireless_gateway(Id, Name, Description)
}
\arguments{
\item{Id}{[required] The ID of the resource to update.}

\item{Name}{The new name of the resource.}

\item{Description}{A new description of the resource.}
}
\description{
Updates properties of a wireless gateway.
}
\section{Request syntax}{
\preformatted{svc$update_wireless_gateway(
  Id = "string",
  Name = "string",
  Description = "string"
)
}
}

\keyword{internal}
