% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_batch_get_channel}
\alias{ivs_batch_get_channel}
\title{Performs GetChannel on multiple ARNs simultaneously}
\usage{
ivs_batch_get_channel(arns)
}
\arguments{
\item{arns}{[required] Array of ARNs, one per channel.}
}
\description{
Performs \code{\link[=ivs_get_channel]{get_channel}} on multiple ARNs
simultaneously.
}
\section{Request syntax}{
\preformatted{svc$batch_get_channel(
  arns = list(
    "string"
  )
)
}
}

\keyword{internal}
