% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_create_stream_key}
\alias{ivs_create_stream_key}
\title{Creates a stream key, used to initiate a stream, for a specified channel
ARN}
\usage{
ivs_create_stream_key(channelArn, tags)
}
\arguments{
\item{channelArn}{[required] ARN of the channel for which to create the stream key.}

\item{tags}{See Channel\\$tags.}
}
\description{
Creates a stream key, used to initiate a stream, for a specified channel
ARN.
}
\details{
Note that CreateChannel creates a stream key. If you subsequently use
CreateStreamKey on the same channel, it will fail because a stream key
already exists and there is a limit of 1 stream key per channel. To
reset the stream key on a channel, use DeleteStreamKey and then
CreateStreamKey.
}
\section{Request syntax}{
\preformatted{svc$create_stream_key(
  channelArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
