% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_delete_channel}
\alias{ivs_delete_channel}
\title{Deletes a specified channel and its associated stream keys}
\usage{
ivs_delete_channel(arn)
}
\arguments{
\item{arn}{[required] ARN of the channel to be deleted.}
}
\description{
Deletes a specified channel and its associated stream keys.
}
\section{Request syntax}{
\preformatted{svc$delete_channel(
  arn = "string"
)
}
}

\keyword{internal}
