% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_get_channel}
\alias{ivs_get_channel}
\title{Gets the channel configuration for the specified channel ARN}
\usage{
ivs_get_channel(arn)
}
\arguments{
\item{arn}{[required] ARN of the channel for which the configuration is to be retrieved.}
}
\description{
Gets the channel configuration for the specified channel ARN. See also
BatchGetChannel.
}
\section{Request syntax}{
\preformatted{svc$get_channel(
  arn = "string"
)
}
}

\keyword{internal}
