% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_get_playback_key_pair}
\alias{ivs_get_playback_key_pair}
\title{Gets a specified playback authorization key pair and returns the arn and
fingerprint}
\usage{
ivs_get_playback_key_pair(arn)
}
\arguments{
\item{arn}{[required] ARN of the key pair to be returned.}
}
\description{
Gets a specified playback authorization key pair and returns the \code{arn}
and \code{fingerprint}. The \code{privateKey} held by the caller can be used to
generate viewer authorization tokens, to grant viewers access to
authorized channels.
}
\section{Request syntax}{
\preformatted{svc$get_playback_key_pair(
  arn = "string"
)
}
}

\keyword{internal}
