% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_get_stream}
\alias{ivs_get_stream}
\title{Gets information about the active (live) stream on a specified channel}
\usage{
ivs_get_stream(channelArn)
}
\arguments{
\item{channelArn}{[required] Channel ARN for stream to be accessed.}
}
\description{
Gets information about the active (live) stream on a specified channel.
}
\section{Request syntax}{
\preformatted{svc$get_stream(
  channelArn = "string"
)
}
}

\keyword{internal}
