% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_get_stream_key}
\alias{ivs_get_stream_key}
\title{Gets stream-key information for a specified ARN}
\usage{
ivs_get_stream_key(arn)
}
\arguments{
\item{arn}{[required] ARN for the stream key to be retrieved.}
}
\description{
Gets stream-key information for a specified ARN.
}
\section{Request syntax}{
\preformatted{svc$get_stream_key(
  arn = "string"
)
}
}

\keyword{internal}
