% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_import_playback_key_pair}
\alias{ivs_import_playback_key_pair}
\title{Imports the public portion of a new key pair and returns its arn and
fingerprint}
\usage{
ivs_import_playback_key_pair(publicKeyMaterial, name, tags)
}
\arguments{
\item{publicKeyMaterial}{[required] The public portion of a customer-generated key pair.}

\item{name}{An arbitrary string (a nickname) assigned to a playback key pair that
helps the customer identify that resource. The value does not need to be
unique.}

\item{tags}{Any tags provided with the request are added to the playback key pair
tags.}
}
\description{
Imports the public portion of a new key pair and returns its \code{arn} and
\code{fingerprint}. The \code{privateKey} can then be used to generate viewer
authorization tokens, to grant viewers access to authorized channels.
}
\section{Request syntax}{
\preformatted{svc$import_playback_key_pair(
  publicKeyMaterial = "string",
  name = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
