% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_channels}
\alias{ivs_list_channels}
\title{Gets summary information about all channels in your account, in the AWS
region where the API request is processed}
\usage{
ivs_list_channels(filterByName, nextToken, maxResults)
}
\arguments{
\item{filterByName}{Filters the channel list to match the specified name.}

\item{nextToken}{The first channel to retrieve. This is used for pagination; see the
\code{nextToken} response field.}

\item{maxResults}{Maximum number of channels to return.}
}
\description{
Gets summary information about all channels in your account, in the AWS
region where the API request is processed. This list can be filtered to
match a specified string.
}
\section{Request syntax}{
\preformatted{svc$list_channels(
  filterByName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
