% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_playback_key_pairs}
\alias{ivs_list_playback_key_pairs}
\title{Gets summary information about playback key pairs}
\usage{
ivs_list_playback_key_pairs(nextToken, maxResults)
}
\arguments{
\item{nextToken}{Maximum number of key pairs to return.}

\item{maxResults}{The first key pair to retrieve. This is used for pagination; see the
\code{nextToken} response field.}
}
\description{
Gets summary information about playback key pairs.
}
\section{Request syntax}{
\preformatted{svc$list_playback_key_pairs(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
