% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_streams}
\alias{ivs_list_streams}
\title{Gets summary information about live streams}
\usage{
ivs_list_streams(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The first stream to retrieve. This is used for pagination; see the
\code{nextToken} response field.}

\item{maxResults}{Maximum number of streams to return.}
}
\description{
Gets summary information about live streams.
}
\section{Request syntax}{
\preformatted{svc$list_streams(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
