% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_stop_stream}
\alias{ivs_stop_stream}
\title{Disconnects the incoming RTMPS stream for the specified channel}
\usage{
ivs_stop_stream(channelArn)
}
\arguments{
\item{channelArn}{[required] ARN of the channel for which the stream is to be stopped.}
}
\description{
Disconnects the incoming RTMPS stream for the specified channel. Can be
used in conjunction with \code{\link[=ivs_delete_stream_key]{delete_stream_key}} to
prevent further streaming to a channel.

Many streaming client-software libraries automatically reconnect a
dropped RTMPS session, so to stop the stream permanently, you may want
to first revoke the \code{streamKey} attached to the channel.
}
\section{Request syntax}{
\preformatted{svc$stop_stream(
  channelArn = "string"
)
}
}

\keyword{internal}
