% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_update_channel}
\alias{ivs_update_channel}
\title{Updates a channel's configuration}
\usage{
ivs_update_channel(arn, name, latencyMode, type)
}
\arguments{
\item{arn}{[required] ARN of the channel to be updated.}

\item{name}{Channel name.}

\item{latencyMode}{Channel latency mode. Default: \code{LOW}.}

\item{type}{Channel type, which determines the allowable resolution and bitrate.
\code{STANDARD}: The stream is transcoded; resolution (width, in landscape
orientation) can be up to 1080p or the input source resolution,
whichever is lower; and bitrate can be up to 8.5 Mbps. \code{BASIC}: The
stream is transfixed; resolution can be up to 480p; and bitrate can be
up to 1.5 Mbps. Default \code{STANDARD}.}
}
\description{
Updates a channel's configuration. This does not affect an ongoing
stream of this channel. You must stop and restart the stream for the
changes to take effect.
}
\section{Request syntax}{
\preformatted{svc$update_channel(
  arn = "string",
  name = "string",
  latencyMode = "NORMAL"|"LOW",
  type = "BASIC"|"STANDARD"
)
}
}

\keyword{internal}
