% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_cluster_operation}
\alias{kafka_describe_cluster_operation}
\title{Returns a description of the cluster operation specified by the ARN}
\usage{
kafka_describe_cluster_operation(ClusterOperationArn)
}
\arguments{
\item{ClusterOperationArn}{[required]
<p>The Amazon Resource Name (ARN) that uniquely identifies the MSK cluster operation.</p>}
}
\description{
<p>Returns a description of the cluster operation specified by the ARN.</p>
}
\section{Request syntax}{
\preformatted{svc$describe_cluster_operation(
  ClusterOperationArn = "string"
)
}
}

\keyword{internal}
