% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_configuration_revision}
\alias{kafka_describe_configuration_revision}
\title{Returns a description of this revision of the configuration}
\usage{
kafka_describe_configuration_revision(Arn, Revision)
}
\arguments{
\item{Arn}{[required]
<p>The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration and all of its revisions.</p>}

\item{Revision}{[required]
<p>A string that uniquely identifies a revision of an MSK configuration.</p>}
}
\description{
\preformatted{        <p>Returns a description of this revision of the configuration.</p>
}
}
\section{Request syntax}{
\preformatted{svc$describe_configuration_revision(
  Arn = "string",
  Revision = 123
)
}
}

\keyword{internal}
