% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_get_compatible_kafka_versions}
\alias{kafka_get_compatible_kafka_versions}
\title{Gets the Apache Kafka versions to which you can update the MSK cluster}
\usage{
kafka_get_compatible_kafka_versions(ClusterArn)
}
\arguments{
\item{ClusterArn}{<p>The Amazon Resource Name (ARN) of the cluster check.</p>}
}
\description{
\preformatted{        <p>Gets the Apache Kafka versions to which you can update the MSK cluster.</p>
}
}
\section{Request syntax}{
\preformatted{svc$get_compatible_kafka_versions(
  ClusterArn = "string"
)
}
}

\keyword{internal}
