% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_list_scram_secrets}
\alias{kafka_list_scram_secrets}
\title{Returns a list of the Scram Secrets associated with an Amazon MSK
cluster}
\usage{
kafka_list_scram_secrets(ClusterArn, MaxResults, NextToken)
}
\arguments{
\item{ClusterArn}{[required]
<p>The arn of the cluster.</p>}

\item{MaxResults}{<p>The maxResults of the query.</p>}

\item{NextToken}{<p>The nextToken of the query.</p>}
}
\description{
\preformatted{        <p>Returns a list of the Scram Secrets associated with an Amazon MSK cluster.</p>
}
}
\section{Request syntax}{
\preformatted{svc$list_scram_secrets(
  ClusterArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
