% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_configuration}
\alias{kafka_update_configuration}
\title{Updates an MSK configuration}
\usage{
kafka_update_configuration(Arn, Description, ServerProperties)
}
\arguments{
\item{Arn}{[required]
<p>The Amazon Resource Name (ARN) of the configuration.</p>}

\item{Description}{<p>The description of the configuration revision.</p>}

\item{ServerProperties}{[required]
<p>Contents of the <filename>server.properties</filename> file. When using the API, you must ensure that the contents of the file are base64 encoded.
When using the AWS Management Console, the SDK, or the AWS CLI, the contents of <filename>server.properties</filename> can be in plaintext.</p>}
}
\description{
\preformatted{        <p>Updates an MSK configuration.</p>
}
}
\section{Request syntax}{
\preformatted{svc$update_configuration(
  Arn = "string",
  Description = "string",
  ServerProperties = raw
)
}
}

\keyword{internal}
