% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_batch_delete_document}
\alias{kendra_batch_delete_document}
\title{Removes one or more documents from an index}
\usage{
kendra_batch_delete_document(IndexId, DocumentIdList,
  DataSourceSyncJobMetricTarget)
}
\arguments{
\item{IndexId}{[required] The identifier of the index that contains the documents to delete.}

\item{DocumentIdList}{[required] One or more identifiers for documents to delete from the index.}

\item{DataSourceSyncJobMetricTarget}{}
}
\description{
Removes one or more documents from an index. The documents must have
been added with the BatchPutDocument operation.
}
\details{
The documents are deleted asynchronously. You can see the progress of
the deletion by using AWS CloudWatch. Any error messages releated to the
processing of the batch are sent to you CloudWatch log.
}
\section{Request syntax}{
\preformatted{svc$batch_delete_document(
  IndexId = "string",
  DocumentIdList = list(
    "string"
  ),
  DataSourceSyncJobMetricTarget = list(
    DataSourceId = "string",
    DataSourceSyncJobId = "string"
  )
)
}
}

\keyword{internal}
