% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_batch_put_document}
\alias{kendra_batch_put_document}
\title{Adds one or more documents to an index}
\usage{
kendra_batch_put_document(IndexId, RoleArn, Documents)
}
\arguments{
\item{IndexId}{[required] The identifier of the index to add the documents to. You need to create
the index first using the CreateIndex operation.}

\item{RoleArn}{The Amazon Resource Name (ARN) of a role that is allowed to run the
\code{BatchPutDocument} operation. For more information, see \href{https://docs.aws.amazon.com/kendra/latest/dg/iam-roles.html}{IAM Roles for Amazon Kendra}.}

\item{Documents}{[required] One or more documents to add to the index.

Documents have the following file size limits.
\itemize{
\item 5 MB total size for inline documents
\item 50 MB total size for files from an S3 bucket
\item 5 MB extracted text for any file
}

For more information about file size and transaction per second quotas,
see \href{https://docs.aws.amazon.com/kendra/latest/dg/quotas.html}{Quotas}.}
}
\description{
Adds one or more documents to an index.
}
\details{
The \code{BatchPutDocument} operation enables you to ingest inline documents
or a set of documents stored in an Amazon S3 bucket. Use this operation
to ingest your text and unstructured text into an index, add custom
attributes to the documents, and to attach an access control list to the
documents added to the index.

The documents are indexed asynchronously. You can see the progress of
the batch using AWS CloudWatch. Any error messages related to processing
the batch are sent to your AWS CloudWatch log.
}
\section{Request syntax}{
\preformatted{svc$batch_put_document(
  IndexId = "string",
  RoleArn = "string",
  Documents = list(
    list(
      Id = "string",
      Title = "string",
      Blob = raw,
      S3Path = list(
        Bucket = "string",
        Key = "string"
      ),
      Attributes = list(
        list(
          Key = "string",
          Value = list(
            StringValue = "string",
            StringListValue = list(
              "string"
            ),
            LongValue = 123,
            DateValue = as.POSIXct(
              "2015-01-01"
            )
          )
        )
      ),
      AccessControlList = list(
        list(
          Name = "string",
          Type = "USER"|"GROUP",
          Access = "ALLOW"|"DENY"
        )
      ),
      ContentType = "PDF"|"HTML"|"MS_WORD"|"PLAIN_TEXT"|"PPT"
    )
  )
)
}
}

\keyword{internal}
