% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_create_faq}
\alias{kendra_create_faq}
\title{Creates an new set of frequently asked question (FAQ) questions and
answers}
\usage{
kendra_create_faq(IndexId, Name, Description, S3Path, RoleArn, Tags)
}
\arguments{
\item{IndexId}{[required] The identifier of the index that contains the FAQ.}

\item{Name}{[required] The name that should be associated with the FAQ.}

\item{Description}{A description of the FAQ.}

\item{S3Path}{[required] The S3 location of the FAQ input data.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of a role with permission to access the
S3 bucket that contains the FAQs. For more information, see \href{https://docs.aws.amazon.com/kendra/latest/dg/iam-roles.html}{IAM Roles for Amazon Kendra}.}

\item{Tags}{A list of key-value pairs that identify the FAQ. You can use the tags to
identify and organize your resources and to control access to resources.}
}
\description{
Creates an new set of frequently asked question (FAQ) questions and
answers.
}
\section{Request syntax}{
\preformatted{svc$create_faq(
  IndexId = "string",
  Name = "string",
  Description = "string",
  S3Path = list(
    Bucket = "string",
    Key = "string"
  ),
  RoleArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
