% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_create_index}
\alias{kendra_create_index}
\title{Creates a new Amazon Kendra index}
\usage{
kendra_create_index(Name, Edition, RoleArn,
  ServerSideEncryptionConfiguration, Description, ClientToken, Tags)
}
\arguments{
\item{Name}{[required] The name for the new index.}

\item{Edition}{The Amazon Kendra edition to use for the index. Choose
\code{DEVELOPER_EDITION} for indexes intended for development, testing, or
proof of concept. Use \code{ENTERPRISE_EDITION} for your production
databases. Once you set the edition for an index, it can\'t be changed.}

\item{RoleArn}{[required] An IAM role that gives Amazon Kendra permissions to access your Amazon
CloudWatch logs and metrics. This is also the role used when you use the
\code{BatchPutDocument} operation to index documents from an Amazon S3
bucket.}

\item{ServerSideEncryptionConfiguration}{The identifier of the AWS KMS customer managed key (CMK) to use to
encrypt data indexed by Amazon Kendra. Amazon Kendra doesn\'t support
asymmetric CMKs.}

\item{Description}{A description for the index.}

\item{ClientToken}{A token that you provide to identify the request to create an index.
Multiple calls to the \code{CreateIndex} operation with the same client token
will create only one index."}

\item{Tags}{A list of key-value pairs that identify the index. You can use the tags
to identify and organize your resources and to control access to
resources.}
}
\description{
Creates a new Amazon Kendra index. Index creation is an asynchronous
operation. To determine if index creation has completed, check the
\code{Status} field returned from a call to . The \code{Status} field is set to
\code{ACTIVE} when the index is ready to use.
}
\details{
Once the index is active you can index your documents using the
operation or using one of the supported data sources.
}
\section{Request syntax}{
\preformatted{svc$create_index(
  Name = "string",
  Edition = "DEVELOPER_EDITION"|"ENTERPRISE_EDITION",
  RoleArn = "string",
  ServerSideEncryptionConfiguration = list(
    KmsKeyId = "string"
  ),
  Description = "string",
  ClientToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
