% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_create_thesaurus}
\alias{kendra_create_thesaurus}
\title{Creates a thesaurus for an index}
\usage{
kendra_create_thesaurus(IndexId, Name, Description, RoleArn, Tags,
  SourceS3Path, ClientToken)
}
\arguments{
\item{IndexId}{[required] The unique identifier of the index for the new thesaurus.}

\item{Name}{[required] The name for the new thesaurus.}

\item{Description}{The description for the new thesaurus.}

\item{RoleArn}{[required] An AWS Identity and Access Management (IAM) role that gives Amazon
Kendra permissions to access thesaurus file specified in \code{SourceS3Path}.}

\item{Tags}{A list of key-value pairs that identify the thesaurus. You can use the
tags to identify and organize your resources and to control access to
resources.}

\item{SourceS3Path}{[required] The thesaurus file Amazon S3 source path.}

\item{ClientToken}{A token that you provide to identify the request to create a thesaurus.
Multiple calls to the \code{CreateThesaurus} operation with the same client
token will create only one index.}
}
\description{
Creates a thesaurus for an index. The thesaurus contains a list of
synonyms in Solr format.
}
\section{Request syntax}{
\preformatted{svc$create_thesaurus(
  IndexId = "string",
  Name = "string",
  Description = "string",
  RoleArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  SourceS3Path = list(
    Bucket = "string",
    Key = "string"
  ),
  ClientToken = "string"
)
}
}

\keyword{internal}
