% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_data_source}
\alias{kendra_delete_data_source}
\title{Deletes an Amazon Kendra data source}
\usage{
kendra_delete_data_source(Id, IndexId)
}
\arguments{
\item{Id}{[required] The unique identifier of the data source to delete.}

\item{IndexId}{[required] The unique identifier of the index associated with the data source.}
}
\description{
Deletes an Amazon Kendra data source. An exception is not thrown if the
data source is already being deleted. While the data source is being
deleted, the \code{Status} field returned by a call to the operation is set
to \code{DELETING}. For more information, see \href{https://docs.aws.amazon.com/kendra/latest/dg/delete-data-source.html}{Deleting Data Sources}.
}
\section{Request syntax}{
\preformatted{svc$delete_data_source(
  Id = "string",
  IndexId = "string"
)
}
}

\keyword{internal}
