% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_faq}
\alias{kendra_delete_faq}
\title{Removes an FAQ from an index}
\usage{
kendra_delete_faq(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the FAQ to remove.}

\item{IndexId}{[required] The index to remove the FAQ from.}
}
\description{
Removes an FAQ from an index.
}
\section{Request syntax}{
\preformatted{svc$delete_faq(
  Id = "string",
  IndexId = "string"
)
}
}

\keyword{internal}
