% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_index}
\alias{kendra_delete_index}
\title{Deletes an existing Amazon Kendra index}
\usage{
kendra_delete_index(Id)
}
\arguments{
\item{Id}{[required] The identifier of the index to delete.}
}
\description{
Deletes an existing Amazon Kendra index. An exception is not thrown if
the index is already being deleted. While the index is being deleted,
the \code{Status} field returned by a call to the DescribeIndex operation is
set to \code{DELETING}.
}
\section{Request syntax}{
\preformatted{svc$delete_index(
  Id = "string"
)
}
}

\keyword{internal}
