% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_thesaurus}
\alias{kendra_delete_thesaurus}
\title{Deletes an existing Amazon Kendra thesaurus}
\usage{
kendra_delete_thesaurus(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the thesaurus to delete.}

\item{IndexId}{[required] The identifier of the index associated with the thesaurus to delete.}
}
\description{
Deletes an existing Amazon Kendra thesaurus.
}
\section{Request syntax}{
\preformatted{svc$delete_thesaurus(
  Id = "string",
  IndexId = "string"
)
}
}

\keyword{internal}
