% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_data_source}
\alias{kendra_describe_data_source}
\title{Gets information about a Amazon Kendra data source}
\usage{
kendra_describe_data_source(Id, IndexId)
}
\arguments{
\item{Id}{[required] The unique identifier of the data source to describe.}

\item{IndexId}{[required] The identifier of the index that contains the data source.}
}
\description{
Gets information about a Amazon Kendra data source.
}
\section{Request syntax}{
\preformatted{svc$describe_data_source(
  Id = "string",
  IndexId = "string"
)
}
}

\keyword{internal}
