% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_faq}
\alias{kendra_describe_faq}
\title{Gets information about an FAQ list}
\usage{
kendra_describe_faq(Id, IndexId)
}
\arguments{
\item{Id}{[required] The unique identifier of the FAQ.}

\item{IndexId}{[required] The identifier of the index that contains the FAQ.}
}
\description{
Gets information about an FAQ list.
}
\section{Request syntax}{
\preformatted{svc$describe_faq(
  Id = "string",
  IndexId = "string"
)
}
}

\keyword{internal}
