% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_index}
\alias{kendra_describe_index}
\title{Describes an existing Amazon Kendra index}
\usage{
kendra_describe_index(Id)
}
\arguments{
\item{Id}{[required] The name of the index to describe.}
}
\description{
Describes an existing Amazon Kendra index
}
\section{Request syntax}{
\preformatted{svc$describe_index(
  Id = "string"
)
}
}

\keyword{internal}
