% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_thesaurus}
\alias{kendra_describe_thesaurus}
\title{Describes an existing Amazon Kendra thesaurus}
\usage{
kendra_describe_thesaurus(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the thesaurus to describe.}

\item{IndexId}{[required] The identifier of the index associated with the thesaurus to describe.}
}
\description{
Describes an existing Amazon Kendra thesaurus.
}
\section{Request syntax}{
\preformatted{svc$describe_thesaurus(
  Id = "string",
  IndexId = "string"
)
}
}

\keyword{internal}
