% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_data_source_sync_jobs}
\alias{kendra_list_data_source_sync_jobs}
\title{Gets statistics about synchronizing Amazon Kendra with a data source}
\usage{
kendra_list_data_source_sync_jobs(Id, IndexId, NextToken, MaxResults,
  StartTimeFilter, StatusFilter)
}
\arguments{
\item{Id}{[required] The identifier of the data source.}

\item{IndexId}{[required] The identifier of the index that contains the data source.}

\item{NextToken}{If the result of the previous request to \code{GetDataSourceSyncJobHistory}
was truncated, include the \code{NextToken} to fetch the next set of jobs.}

\item{MaxResults}{The maximum number of synchronization jobs to return in the response. If
there are fewer results in the list, this response contains only the
actual results.}

\item{StartTimeFilter}{When specified, the synchronization jobs returned in the list are
limited to jobs between the specified dates.}

\item{StatusFilter}{When specified, only returns synchronization jobs with the \code{Status}
field equal to the specified status.}
}
\description{
Gets statistics about synchronizing Amazon Kendra with a data source.
}
\section{Request syntax}{
\preformatted{svc$list_data_source_sync_jobs(
  Id = "string",
  IndexId = "string",
  NextToken = "string",
  MaxResults = 123,
  StartTimeFilter = list(
    StartTime = as.POSIXct(
      "2015-01-01"
    ),
    EndTime = as.POSIXct(
      "2015-01-01"
    )
  ),
  StatusFilter = "FAILED"|"SUCCEEDED"|"SYNCING"|"INCOMPLETE"|"STOPPING"|"ABORTED"|"SYNCING_INDEXING"
)
}
}

\keyword{internal}
