% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_data_sources}
\alias{kendra_list_data_sources}
\title{Lists the data sources that you have created}
\usage{
kendra_list_data_sources(IndexId, NextToken, MaxResults)
}
\arguments{
\item{IndexId}{[required] The identifier of the index that contains the data source.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of data sources
(\code{DataSourceSummaryItems}).}

\item{MaxResults}{The maximum number of data sources to return.}
}
\description{
Lists the data sources that you have created.
}
\section{Request syntax}{
\preformatted{svc$list_data_sources(
  IndexId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
