% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_faqs}
\alias{kendra_list_faqs}
\title{Gets a list of FAQ lists associated with an index}
\usage{
kendra_list_faqs(IndexId, NextToken, MaxResults)
}
\arguments{
\item{IndexId}{[required] The index that contains the FAQ lists.}

\item{NextToken}{If the result of the previous request to \code{\link[=kendra_list_faqs]{list_faqs}}
was truncated, include the \code{NextToken} to fetch the next set of FAQs.}

\item{MaxResults}{The maximum number of FAQs to return in the response. If there are fewer
results in the list, this response contains only the actual results.}
}
\description{
Gets a list of FAQ lists associated with an index.
}
\section{Request syntax}{
\preformatted{svc$list_faqs(
  IndexId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
