% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_indices}
\alias{kendra_list_indices}
\title{Lists the Amazon Kendra indexes that you have created}
\usage{
kendra_list_indices(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of indexes
(\code{DataSourceSummaryItems}).}

\item{MaxResults}{The maximum number of data sources to return.}
}
\description{
Lists the Amazon Kendra indexes that you have created.
}
\section{Request syntax}{
\preformatted{svc$list_indices(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
