% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_tags_for_resource}
\alias{kendra_list_tags_for_resource}
\title{Gets a list of tags associated with a specified resource}
\usage{
kendra_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the index, FAQ, or data source to get
a list of tags for.}
}
\description{
Gets a list of tags associated with a specified resource. Indexes, FAQs,
and data sources can have tags associated with them.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}
}

\keyword{internal}
