% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_thesauri}
\alias{kendra_list_thesauri}
\title{Lists the Amazon Kendra thesauri associated with an index}
\usage{
kendra_list_thesauri(IndexId, NextToken, MaxResults)
}
\arguments{
\item{IndexId}{[required] The identifier of the index associated with the thesaurus to list.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of thesauri
(\code{ThesaurusSummaryItems}).}

\item{MaxResults}{The maximum number of thesauri to return.}
}
\description{
Lists the Amazon Kendra thesauri associated with an index.
}
\section{Request syntax}{
\preformatted{svc$list_thesauri(
  IndexId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
