% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_stop_data_source_sync_job}
\alias{kendra_stop_data_source_sync_job}
\title{Stops a running synchronization job}
\usage{
kendra_stop_data_source_sync_job(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the data source for which to stop the synchronization
jobs.}

\item{IndexId}{[required] The identifier of the index that contains the data source.}
}
\description{
Stops a running synchronization job. You can\'t stop a scheduled
synchronization job.
}
\section{Request syntax}{
\preformatted{svc$stop_data_source_sync_job(
  Id = "string",
  IndexId = "string"
)
}
}

\keyword{internal}
