% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_untag_resource}
\alias{kendra_untag_resource}
\title{Removes a tag from an index, FAQ, or a data source}
\usage{
kendra_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the index, FAQ, or data source to
remove the tag from.}

\item{TagKeys}{[required] A list of tag keys to remove from the index, FAQ, or data source. If a
tag key does not exist on the resource, it is ignored.}
}
\description{
Removes a tag from an index, FAQ, or a data source.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
