% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_update_index}
\alias{kendra_update_index}
\title{Updates an existing Amazon Kendra index}
\usage{
kendra_update_index(Id, Name, RoleArn, Description,
  DocumentMetadataConfigurationUpdates, CapacityUnits)
}
\arguments{
\item{Id}{[required] The identifier of the index to update.}

\item{Name}{The name of the index to update.}

\item{RoleArn}{A new IAM role that gives Amazon Kendra permission to access your Amazon
CloudWatch logs.}

\item{Description}{A new description for the index.}

\item{DocumentMetadataConfigurationUpdates}{The document metadata to update.}

\item{CapacityUnits}{Sets the number of addtional storage and query capacity units that
should be used by the index. You can change the capacity of the index up
to 5 times per day.

If you are using extra storage units, you can\'t reduce the storage
capacity below that required to meet the storage needs for your index.}
}
\description{
Updates an existing Amazon Kendra index.
}
\section{Request syntax}{
\preformatted{svc$update_index(
  Id = "string",
  Name = "string",
  RoleArn = "string",
  Description = "string",
  DocumentMetadataConfigurationUpdates = list(
    list(
      Name = "string",
      Type = "STRING_VALUE"|"STRING_LIST_VALUE"|"LONG_VALUE"|"DATE_VALUE",
      Relevance = list(
        Freshness = TRUE|FALSE,
        Importance = 123,
        Duration = "string",
        RankOrder = "ASCENDING"|"DESCENDING",
        ValueImportanceMap = list(
          123
        )
      ),
      Search = list(
        Facetable = TRUE|FALSE,
        Searchable = TRUE|FALSE,
        Displayable = TRUE|FALSE
      )
    )
  ),
  CapacityUnits = list(
    StorageCapacityUnits = 123,
    QueryCapacityUnits = 123
  )
)
}
}

\keyword{internal}
