% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_update_thesaurus}
\alias{kendra_update_thesaurus}
\title{Updates a thesaurus file associated with an index}
\usage{
kendra_update_thesaurus(Id, Name, IndexId, Description, RoleArn,
  SourceS3Path)
}
\arguments{
\item{Id}{[required] The identifier of the thesaurus to update.}

\item{Name}{The updated name of the thesaurus.}

\item{IndexId}{[required] The identifier of the index associated with the thesaurus to update.}

\item{Description}{The updated description of the thesaurus.}

\item{RoleArn}{The updated role ARN of the thesaurus.}

\item{SourceS3Path}{}
}
\description{
Updates a thesaurus file associated with an index.
}
\section{Request syntax}{
\preformatted{svc$update_thesaurus(
  Id = "string",
  Name = "string",
  IndexId = "string",
  Description = "string",
  RoleArn = "string",
  SourceS3Path = list(
    Bucket = "string",
    Key = "string"
  )
)
}
}

\keyword{internal}
