% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_create_application_snapshot}
\alias{kinesisanalyticsv2_create_application_snapshot}
\title{Creates a snapshot of the application's state data}
\usage{
kinesisanalyticsv2_create_application_snapshot(ApplicationName,
  SnapshotName)
}
\arguments{
\item{ApplicationName}{[required] The name of an existing application}

\item{SnapshotName}{[required] An identifier for the application snapshot.}
}
\description{
Creates a snapshot of the application's state data.
}
\section{Request syntax}{
\preformatted{svc$create_application_snapshot(
  ApplicationName = "string",
  SnapshotName = "string"
)
}
}

\keyword{internal}
