% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_stop_application}
\alias{kinesisanalyticsv2_stop_application}
\title{Stops the application from processing data}
\usage{
kinesisanalyticsv2_stop_application(ApplicationName)
}
\arguments{
\item{ApplicationName}{[required] The name of the running application to stop.}
}
\description{
Stops the application from processing data. You can stop an application
only if it is in the running state. You can use the DescribeApplication
operation to find the application state.
}
\section{Request syntax}{
\preformatted{kinesisanalyticsv2$stop_application(
  ApplicationName = "string"
)
}
}

\keyword{internal}
