% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_delete_stream}
\alias{kinesisvideo_delete_stream}
\title{Deletes a Kinesis video stream and the data contained in the stream}
\usage{
kinesisvideo_delete_stream(StreamARN, CurrentVersion)
}
\arguments{
\item{StreamARN}{[required] The Amazon Resource Name (ARN) of the stream that you want to delete.}

\item{CurrentVersion}{Optional: The version of the stream that you want to delete.

Specify the version as a safeguard to ensure that your are deleting the
correct stream. To get the stream version, use the \code{DescribeStream} API.

If not specified, only the \code{CreationTime} is checked before deleting the
stream.}
}
\description{
Deletes a Kinesis video stream and the data contained in the stream.

This method marks the stream for deletion, and makes the data in the
stream inaccessible immediately.

To ensure that you have the latest version of the stream before deleting
it, you can specify the stream version. Kinesis Video Streams assigns a
version to each stream. When you update a stream, Kinesis Video Streams
assigns a new version number. To get the latest stream version, use the
\code{DescribeStream} API.

This operation requires permission for the \code{KinesisVideo:DeleteStream}
action.
}
\section{Request syntax}{
\preformatted{svc$delete_stream(
  StreamARN = "string",
  CurrentVersion = "string"
)
}
}

\keyword{internal}
