% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_untag_resource}
\alias{kinesisvideo_untag_resource}
\title{Removes one or more tags from a signaling channel}
\usage{
kinesisvideo_untag_resource(ResourceARN, TagKeyList)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the signaling channel from which you want to remove tags.}

\item{TagKeyList}{[required] A list of the keys of the tags that you want to remove.}
}
\description{
Removes one or more tags from a signaling channel. In the request,
specify only a tag key or keys; don\'t specify the value. If you specify
a tag key that does not exist, it\'s ignored.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeyList = list(
    "string"
  )
)
}
}

\keyword{internal}
