% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_untag_stream}
\alias{kinesisvideo_untag_stream}
\title{Removes one or more tags from a stream}
\usage{
kinesisvideo_untag_stream(StreamARN, StreamName, TagKeyList)
}
\arguments{
\item{StreamARN}{The Amazon Resource Name (ARN) of the stream that you want to remove
tags from.}

\item{StreamName}{The name of the stream that you want to remove tags from.}

\item{TagKeyList}{[required] A list of the keys of the tags that you want to remove.}
}
\description{
Removes one or more tags from a stream. In the request, specify only a
tag key or keys; don't specify the value. If you specify a tag key that
does not exist, it's ignored.

In the request, you must provide the \code{StreamName} or \code{StreamARN}.
}
\section{Request syntax}{
\preformatted{svc$untag_stream(
  StreamARN = "string",
  StreamName = "string",
  TagKeyList = list(
    "string"
  )
)
}
}

\keyword{internal}
