% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideoarchivedmedia_service.R
\name{kinesisvideoarchivedmedia}
\alias{kinesisvideoarchivedmedia}
\title{Amazon Kinesis Video Streams Archived Media}
\usage{
kinesisvideoarchivedmedia()
}
\description{
Amazon Kinesis Video Streams Archived Media
}
\section{Example}{
\preformatted{kinesisvideoarchivedmedia <- paws::kinesisvideoarchivedmedia()
kinesisvideoarchivedmedia$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=kinesisvideoarchivedmedia_get_hls_streaming_session_url]{get_hls_streaming_session_url} \tab Retrieves an HTTP Live Streaming (HLS) URL for the stream \cr
\link[=kinesisvideoarchivedmedia_get_media_for_fragment_list]{get_media_for_fragment_list} \tab Gets media for a list of fragments (specified by fragment number) from the archived data in an Amazon Kinesis video stream\cr
\link[=kinesisvideoarchivedmedia_list_fragments]{list_fragments} \tab Returns a list of Fragment objects from the specified stream and start location within the archived data
}
}

