% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideoarchivedmedia_operations.R
\name{kinesisvideoarchivedmedia_list_fragments}
\alias{kinesisvideoarchivedmedia_list_fragments}
\title{Returns a list of Fragment objects from the specified stream and start
location within the archived data}
\usage{
kinesisvideoarchivedmedia_list_fragments(StreamName, MaxResults,
  NextToken, FragmentSelector)
}
\arguments{
\item{StreamName}{[required] The name of the stream from which to retrieve a fragment list.}

\item{MaxResults}{The total number of fragments to return. If the total number of
fragments available is more than the value specified in \code{max-results},
then a ListFragmentsOutput\$NextToken is provided in the output that you
can use to resume pagination.}

\item{NextToken}{A token to specify where to start paginating. This is the
ListFragmentsOutput\$NextToken from a previously truncated response.}

\item{FragmentSelector}{Describes the time stamp range and time stamp origin for the range of
fragments to return.}
}
\description{
Returns a list of Fragment objects from the specified stream and start
location within the archived data.
}
\section{Request syntax}{
\preformatted{kinesisvideoarchivedmedia$list_fragments(
  StreamName = "string",
  MaxResults = 123,
  NextToken = "string",
  FragmentSelector = list(
    FragmentSelectorType = "PRODUCER_TIMESTAMP"|"SERVER_TIMESTAMP",
    TimestampRange = list(
      StartTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      EndTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}

\keyword{internal}
