% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideomedia_service.R
\name{kinesisvideomedia}
\alias{kinesisvideomedia}
\title{Amazon Kinesis Video Streams Media}
\usage{
kinesisvideomedia(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Kinesis Video Streams Media
}
\section{Service syntax}{
\preformatted{svc <- kinesisvideomedia(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=kinesisvideomedia_get_media]{get_media} \tab Use this API to retrieve media content from a Kinesis video stream
}
}

\examples{
\donttest{svc <- kinesisvideomedia()
svc$get_media(
  Foo = 123
)}

}
