% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideosignalingchannels_service.R
\name{kinesisvideosignalingchannels}
\alias{kinesisvideosignalingchannels}
\title{Amazon Kinesis Video Signaling Channels}
\usage{
kinesisvideosignalingchannels(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Kinesis Video Streams Signaling Service is a intermediate service that
establishes a communication channel for discovering peers, transmitting
offers and answers in order to establish peer-to-peer connection in
webRTC technology.
}
\section{Service syntax}{
\preformatted{svc <- kinesisvideosignalingchannels(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=kinesisvideosignalingchannels_get_ice_server_config]{get_ice_server_config} \tab Gets the Interactive Connectivity Establishment (ICE) server configuration information, including URIs, username, and password which can be used to configure the WebRTC connection\cr
\link[=kinesisvideosignalingchannels_send_alexa_offer_to_master]{send_alexa_offer_to_master} \tab This API allows you to connect WebRTC-enabled devices with Alexa display devices
}
}

\examples{
\dontrun{
svc <- kinesisvideosignalingchannels()
svc$get_ice_server_config(
  Foo = 123
)
}

}
