% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideosignalingchannels_operations.R
\name{kinesisvideosignalingchannels_send_alexa_offer_to_master}
\alias{kinesisvideosignalingchannels_send_alexa_offer_to_master}
\title{This API allows you to connect WebRTC-enabled devices with Alexa display
devices}
\usage{
kinesisvideosignalingchannels_send_alexa_offer_to_master(ChannelARN,
  SenderClientId, MessagePayload)
}
\arguments{
\item{ChannelARN}{[required] The ARN of the signaling channel by which Alexa and the master peer
communicate.}

\item{SenderClientId}{[required] The unique identifier for the sender client.}

\item{MessagePayload}{[required] The base64-encoded SDP offer content.}
}
\description{
This API allows you to connect WebRTC-enabled devices with Alexa display
devices. When invoked, it sends the Alexa Session Description Protocol
(SDP) offer to the master peer. The offer is delivered as soon as the
master is connected to the specified signaling channel. This API returns
the SDP answer from the connected master. If the master is not connected
to the signaling channel, redelivery requests are made until the message
expires.
}
\section{Request syntax}{
\preformatted{svc$send_alexa_offer_to_master(
  ChannelARN = "string",
  SenderClientId = "string",
  MessagePayload = "string"
)
}
}

\keyword{internal}
