% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_disable_key}
\alias{kms_disable_key}
\title{Sets the state of a customer master key (CMK) to disabled}
\usage{
kms_disable_key(KeyId)
}
\arguments{
\item{KeyId}{[required] A unique identifier for the customer master key (CMK).

Specify the key ID or the Amazon Resource Name (ARN) of the CMK.

For example:
\itemize{
\item Key ID: \verb{1234abcd-12ab-34cd-56ef-1234567890ab}
\item Key ARN:
\verb{arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab}
}

To get the key ID and key ARN for a CMK, use
\href{kms_list_keys}{\code{list_keys}} or \code{\link[=kms_describe_key]{describe_key}}.}
}
\description{
Sets the state of a customer master key (CMK) to disabled. This change
temporarily prevents use of the CMK for \href{https://docs.aws.amazon.com/kms/latest/developerguide/concepts.html#cryptographic-operations}{cryptographic operations}.

For more information about how key state affects the use of a CMK, see
\href{https://docs.aws.amazon.com/kms/latest/developerguide/key-state.html}{How Key State Affects the Use of a Customer Master Key}
in the \strong{AWS Key Management Service Developer Guide} .

The CMK that you use for this operation must be in a compatible key
state. For details, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/key-state.html}{How Key State Affects Use of a Customer Master Key}
in the \emph{AWS Key Management Service Developer Guide}.

\strong{Cross-account use}: No. You cannot perform this operation on a CMK in
a different AWS account.

\strong{Required permissions}:
\href{https://docs.aws.amazon.com/kms/latest/developerguide/kms-api-permissions-reference.html}{kms:DisableKey}
(key policy)

\strong{Related operations}: \code{\link[=kms_enable_key]{enable_key}}
}
\section{Request syntax}{
\preformatted{svc$disable_key(
  KeyId = "string"
)
}
}

\examples{
\dontrun{
# The following example disables the specified CMK.
svc$disable_key(
  KeyId = "1234abcd-12ab-34cd-56ef-1234567890ab"
)
}

}
\keyword{internal}
