% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_batch_grant_permissions}
\alias{lakeformation_batch_grant_permissions}
\title{Batch operation to grant permissions to the principal}
\usage{
lakeformation_batch_grant_permissions(CatalogId, Entries)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your AWS Lake Formation environment.}

\item{Entries}{[required] A list of up to 20 entries for resource permissions to be granted by
batch operation to the principal.}
}
\description{
Batch operation to grant permissions to the principal.
}
\section{Request syntax}{
\preformatted{svc$batch_grant_permissions(
  CatalogId = "string",
  Entries = list(
    list(
      Id = "string",
      Principal = list(
        DataLakePrincipalIdentifier = "string"
      ),
      Resource = list(
        Catalog = list(),
        Database = list(
          Name = "string"
        ),
        Table = list(
          DatabaseName = "string",
          Name = "string"
        ),
        TableWithColumns = list(
          DatabaseName = "string",
          Name = "string",
          ColumnNames = list(
            "string"
          ),
          ColumnWildcard = list(
            ExcludedColumnNames = list(
              "string"
            )
          )
        ),
        DataLocation = list(
          ResourceArn = "string"
        )
      ),
      Permissions = list(
        "ALL"|"SELECT"|"ALTER"|"DROP"|"DELETE"|"INSERT"|"CREATE_DATABASE"|"CREATE_TABLE"|"DATA_LOCATION_ACCESS"
      ),
      PermissionsWithGrantOption = list(
        "ALL"|"SELECT"|"ALTER"|"DROP"|"DELETE"|"INSERT"|"CREATE_DATABASE"|"CREATE_TABLE"|"DATA_LOCATION_ACCESS"
      )
    )
  )
)
}
}

\keyword{internal}
