% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_describe_resource}
\alias{lakeformation_describe_resource}
\title{Retrieves the current data access role for the given resource registered
in AWS Lake Formation}
\usage{
lakeformation_describe_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The resource ARN.}
}
\description{
Retrieves the current data access role for the given resource registered
in AWS Lake Formation.
}
\section{Request syntax}{
\preformatted{svc$describe_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
