% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_list_permissions}
\alias{lakeformation_list_permissions}
\title{Returns a list of the principal permissions on the resource, filtered by
the permissions of the caller}
\usage{
lakeformation_list_permissions(CatalogId, Principal, ResourceType,
  Resource, NextToken, MaxResults)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your AWS Lake Formation environment.}

\item{Principal}{Specifies a principal to filter the permissions returned.}

\item{ResourceType}{Specifies a resource type to filter the permissions returned.}

\item{Resource}{A resource where you will get a list of the principal permissions.

This operation does not support getting privileges on a table with
columns. Instead, call this operation on the table, and the operation
returns the table and the table w columns.}

\item{NextToken}{A continuation token, if this is not the first call to retrieve this
list.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Returns a list of the principal permissions on the resource, filtered by
the permissions of the caller. For example, if you are granted an ALTER
permission, you are able to see only the principal permissions for
ALTER.

This operation returns only those permissions that have been explicitly
granted.

For information about permissions, see Security and Access Control to
Metadata and Data.
}
\section{Request syntax}{
\preformatted{svc$list_permissions(
  CatalogId = "string",
  Principal = list(
    DataLakePrincipalIdentifier = "string"
  ),
  ResourceType = "CATALOG"|"DATABASE"|"TABLE"|"DATA_LOCATION",
  Resource = list(
    Catalog = list(),
    Database = list(
      CatalogId = "string",
      Name = "string"
    ),
    Table = list(
      CatalogId = "string",
      DatabaseName = "string",
      Name = "string",
      TableWildcard = list()
    ),
    TableWithColumns = list(
      CatalogId = "string",
      DatabaseName = "string",
      Name = "string",
      ColumnNames = list(
        "string"
      ),
      ColumnWildcard = list(
        ExcludedColumnNames = list(
          "string"
        )
      )
    ),
    DataLocation = list(
      CatalogId = "string",
      ResourceArn = "string"
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
