% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_put_data_lake_settings}
\alias{lakeformation_put_data_lake_settings}
\title{The AWS Lake Formation principal}
\usage{
lakeformation_put_data_lake_settings(CatalogId, DataLakeSettings)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your AWS Lake Formation environment.}

\item{DataLakeSettings}{[required] A list of AWS Lake Formation principals.}
}
\description{
The AWS Lake Formation principal.
}
\section{Request syntax}{
\preformatted{svc$put_data_lake_settings(
  CatalogId = "string",
  DataLakeSettings = list(
    DataLakeAdmins = list(
      list(
        DataLakePrincipalIdentifier = "string"
      )
    ),
    CreateDatabaseDefaultPermissions = list(
      list(
        Principal = list(
          DataLakePrincipalIdentifier = "string"
        ),
        Permissions = list(
          "ALL"|"SELECT"|"ALTER"|"DROP"|"DELETE"|"INSERT"|"CREATE_DATABASE"|"CREATE_TABLE"|"DATA_LOCATION_ACCESS"
        )
      )
    ),
    CreateTableDefaultPermissions = list(
      list(
        Principal = list(
          DataLakePrincipalIdentifier = "string"
        ),
        Permissions = list(
          "ALL"|"SELECT"|"ALTER"|"DROP"|"DELETE"|"INSERT"|"CREATE_DATABASE"|"CREATE_TABLE"|"DATA_LOCATION_ACCESS"
        )
      )
    )
  )
)
}
}

\keyword{internal}
