% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_register_resource}
\alias{lakeformation_register_resource}
\title{Registers the resource as managed by the Data Catalog}
\usage{
lakeformation_register_resource(ResourceArn, UseServiceLinkedRole,
  RoleArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to
register.}

\item{UseServiceLinkedRole}{Designates a trusted caller, an IAM principal, by registering this
caller with the Data Catalog.}

\item{RoleArn}{The identifier for the role.}
}
\description{
Registers the resource as managed by the Data Catalog.
}
\details{
To add or update data, Lake Formation needs read/write access to the
chosen Amazon S3 path. Choose a role that you know has permission to do
this, or choose the AWSServiceRoleForLakeFormationDataAccess
service-linked role. When you register the first Amazon S3 path, the
service-linked role and a new inline policy are created on your behalf.
Lake Formation adds the first path to the inline policy and attaches it
to the service-linked role. When you register subsequent paths, Lake
Formation adds the path to the existing policy.
}
\section{Request syntax}{
\preformatted{svc$register_resource(
  ResourceArn = "string",
  UseServiceLinkedRole = TRUE|FALSE,
  RoleArn = "string"
)
}
}

\keyword{internal}
