% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_delete_alias}
\alias{lambda_delete_alias}
\title{Deletes a Lambda function alias}
\usage{
lambda_delete_alias(FunctionName, Name)
}
\arguments{
\item{FunctionName}{[required] The name of the Lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} -
\code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only
the function name, it is limited to 64 characters in length.}

\item{Name}{[required] The name of the alias.}
}
\description{
Deletes a Lambda function
\href{https://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{alias}.
}
\section{Request syntax}{
\preformatted{lambda$delete_alias(
  FunctionName = "string",
  Name = "string"
)
}
}

\examples{
# This operation deletes a Lambda function alias
\donttest{lambda$delete_alias(
  FunctionName = "myFunction",
  Name = "alias"
)}

}
\keyword{internal}
