% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_delete_function_concurrency}
\alias{lambda_delete_function_concurrency}
\title{Removes a concurrent execution limit from a function}
\usage{
lambda_delete_function_concurrency(FunctionName)
}
\arguments{
\item{FunctionName}{[required] The name of the Lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{my-function}.
\item \strong{Function ARN} -
\code{arn:aws:lambda:us-west-2:123456789012:function:my-function}.
\item \strong{Partial ARN} - \code{123456789012:function:my-function}.
}

The length constraint applies only to the full ARN. If you specify only
the function name, it is limited to 64 characters in length.}
}
\description{
Removes a concurrent execution limit from a function.
}
\section{Request syntax}{
\preformatted{svc$delete_function_concurrency(
  FunctionName = "string"
)
}
}

\keyword{internal}
