% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_account_settings}
\alias{lambda_get_account_settings}
\title{Retrieves details about your account's limits and usage in a region}
\usage{
lambda_get_account_settings()
}
\description{
Retrieves details about your account's
\href{http://docs.aws.amazon.com/lambda/latest/dg/limits.html}{limits} and
usage in a region.
}
\section{Request syntax}{
\preformatted{lambda$get_account_settings()
}
}

\examples{
# This operation retrieves a Lambda customer's account settings
\donttest{lambda$get_account_settings()}

}
\keyword{internal}
