% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_tag_resource}
\alias{lambda_tag_resource}
\title{Adds tags to a function}
\usage{
lambda_tag_resource(Resource, Tags)
}
\arguments{
\item{Resource}{[required] The function's Amazon Resource Name (ARN).}

\item{Tags}{[required] A list of tags to apply to the function.}
}
\description{
Adds \href{http://docs.aws.amazon.com/lambda/latest/dg/tagging.html}{tags} to
a function.
}
\section{Request syntax}{
\preformatted{lambda$tag_resource(
  Resource = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
