% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_bot_alias}
\alias{lexmodelbuildingservice_get_bot_alias}
\title{Returns information about an Amazon Lex bot alias}
\usage{
lexmodelbuildingservice_get_bot_alias(name, botName)
}
\arguments{
\item{name}{[required] The name of the bot alias. The name is case sensitive.}

\item{botName}{[required] The name of the bot.}
}
\description{
Returns information about an Amazon Lex bot alias. For more information
about aliases, see versioning-aliases.
}
\details{
This operation requires permissions for the \code{lex:GetBotAlias} action.
}
\section{Request syntax}{
\preformatted{svc$get_bot_alias(
  name = "string",
  botName = "string"
)
}
}

\keyword{internal}
