% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_untag_resource}
\alias{lexmodelbuildingservice_untag_resource}
\title{Removes tags from a bot, bot alias or bot channel}
\usage{
lexmodelbuildingservice_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to remove the tags from.}

\item{tagKeys}{[required] A list of tag keys to remove from the resource. If a tag key does not
exist on the resource, it is ignored.}
}
\description{
Removes tags from a bot, bot alias or bot channel.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
