% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_delete_grant}
\alias{licensemanager_delete_grant}
\title{Deletes the specified grant}
\usage{
licensemanager_delete_grant(GrantArn, Version)
}
\arguments{
\item{GrantArn}{[required] Amazon Resource Name (ARN) of the grant.}

\item{Version}{[required] Current version of the grant.}
}
\description{
Deletes the specified grant.
}
\section{Request syntax}{
\preformatted{svc$delete_grant(
  GrantArn = "string",
  Version = "string"
)
}
}

\keyword{internal}
