% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_delete_license_configuration}
\alias{licensemanager_delete_license_configuration}
\title{Deletes an existing license configuration}
\usage{
licensemanager_delete_license_configuration(LicenseConfigurationArn)
}
\arguments{
\item{LicenseConfigurationArn}{[required] Unique ID of the configuration object to delete.}
}
\description{
Deletes an existing license configuration. This action fails if the
configuration is in use.
}
\section{Request syntax}{
\preformatted{licensemanager$delete_license_configuration(
  LicenseConfigurationArn = "string"
)
}
}

\keyword{internal}
