% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_license}
\alias{licensemanager_get_license}
\title{Gets detailed information about the specified license}
\usage{
licensemanager_get_license(LicenseArn, Version)
}
\arguments{
\item{LicenseArn}{[required] Amazon Resource Name (ARN) of the license.}

\item{Version}{License version.}
}
\description{
Gets detailed information about the specified license.
}
\section{Request syntax}{
\preformatted{svc$get_license(
  LicenseArn = "string",
  Version = "string"
)
}
}

\keyword{internal}
