% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_license_configuration}
\alias{licensemanager_get_license_configuration}
\title{Gets detailed information about the specified license configuration}
\usage{
licensemanager_get_license_configuration(LicenseConfigurationArn)
}
\arguments{
\item{LicenseConfigurationArn}{[required] Amazon Resource Name (ARN) of the license configuration.}
}
\description{
Gets detailed information about the specified license configuration.
}
\section{Request syntax}{
\preformatted{svc$get_license_configuration(
  LicenseConfigurationArn = "string"
)
}
}

\keyword{internal}
