% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_service_settings}
\alias{licensemanager_get_service_settings}
\title{Gets License Manager settings for a region}
\usage{
licensemanager_get_service_settings()
}
\description{
Gets License Manager settings for a region. Exposes the configured S3
bucket, SNS topic, etc., for inspection.
}
\section{Request syntax}{
\preformatted{licensemanager$get_service_settings()
}
}

\keyword{internal}
