% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_tags_for_resource}
\alias{licensemanager_list_tags_for_resource}
\title{Lists tags attached to a resource}
\usage{
licensemanager_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] ARN for the resource.}
}
\description{
Lists tags attached to a resource.
}
\section{Request syntax}{
\preformatted{licensemanager$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
