% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_tag_resource}
\alias{licensemanager_tag_resource}
\title{Attach one of more tags to any resource}
\usage{
licensemanager_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] Resource of the ARN to be tagged.}

\item{Tags}{[required] Names of the tags to attach to the resource.}
}
\description{
Attach one of more tags to any resource.
}
\section{Request syntax}{
\preformatted{licensemanager$tag_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
