% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_untag_resource}
\alias{licensemanager_untag_resource}
\title{Remove tags from a resource}
\usage{
licensemanager_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] ARN of the resource.}

\item{TagKeys}{[required] List keys identifying tags to remove.}
}
\description{
Remove tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
