% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_alarm}
\alias{lightsail_delete_alarm}
\title{Deletes an alarm}
\usage{
lightsail_delete_alarm(alarmName)
}
\arguments{
\item{alarmName}{[required] The name of the alarm to delete.}
}
\description{
Deletes an alarm.
}
\details{
An alarm is used to monitor a single metric for one of your resources.
When a metric condition is met, the alarm can notify you by email, SMS
text message, and a banner displayed on the Amazon Lightsail console.
For more information, see \href{https://lightsail.aws.amazon.com/ls/docs/en_us/articles/amazon-lightsail-alarms}{Alarms in Amazon Lightsail}.
}
\section{Request syntax}{
\preformatted{svc$delete_alarm(
  alarmName = "string"
)
}
}

\keyword{internal}
