% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_export_snapshot_records}
\alias{lightsail_get_export_snapshot_records}
\title{Returns the export snapshot record created as a result of the export
snapshot operation}
\usage{
lightsail_get_export_snapshot_records(pageToken)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_export_snapshot_records]{get_export_snapshot_records}}
request. If your results are paginated, the response will return a next
page token that you can specify as the page token in a subsequent
request.}
}
\description{
Returns the export snapshot record created as a result of the
\verb{export snapshot} operation.

An export snapshot record can be used to create a new Amazon EC2
instance and its related resources with the
\verb{create cloud formation stack} operation.
}
\section{Request syntax}{
\preformatted{svc$get_export_snapshot_records(
  pageToken = "string"
)
}
}

\keyword{internal}
