% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database_bundles}
\alias{lightsail_get_relational_database_bundles}
\title{Returns the list of bundles that are available in Amazon Lightsail}
\usage{
lightsail_get_relational_database_bundles(pageToken)
}
\arguments{
\item{pageToken}{A token used for advancing to a specific page of results for your
\verb{get relational database bundles} request.}
}
\description{
Returns the list of bundles that are available in Amazon Lightsail. A
bundle describes the performance specifications for a database.
}
\details{
You can use a bundle ID to create a new database with explicit
performance specifications.
}
\section{Request syntax}{
\preformatted{svc$get_relational_database_bundles(
  pageToken = "string"
)
}
}

\keyword{internal}
