% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database_snapshots}
\alias{lightsail_get_relational_database_snapshots}
\title{Returns information about all of your database snapshots in Amazon
Lightsail}
\usage{
lightsail_get_relational_database_snapshots(pageToken)
}
\arguments{
\item{pageToken}{A token used for advancing to a specific page of results for your
\code{get relational database snapshots} request.}
}
\description{
Returns information about all of your database snapshots in Amazon
Lightsail.
}
\section{Request syntax}{
\preformatted{lightsail$get_relational_database_snapshots(
  pageToken = "string"
)
}
}

\keyword{internal}
