% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_databases}
\alias{lightsail_get_relational_databases}
\title{Returns information about all of your databases in Amazon Lightsail}
\usage{
lightsail_get_relational_databases(pageToken)
}
\arguments{
\item{pageToken}{A token used for advancing to a specific page of results for your
\code{get relational database} request.}
}
\description{
Returns information about all of your databases in Amazon Lightsail.
}
\section{Request syntax}{
\preformatted{lightsail$get_relational_databases(
  pageToken = "string"
)
}
}

\keyword{internal}
