% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_reboot_relational_database}
\alias{lightsail_reboot_relational_database}
\title{Restarts a specific database in Amazon Lightsail}
\usage{
lightsail_reboot_relational_database(relationalDatabaseName)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database to reboot.}
}
\description{
Restarts a specific database in Amazon Lightsail.
}
\details{
The \verb{reboot relational database} operation supports tag-based access
control via resource tags applied to the resource identified by
relationalDatabaseName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en_us/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\section{Request syntax}{
\preformatted{svc$reboot_relational_database(
  relationalDatabaseName = "string"
)
}
}

\keyword{internal}
