% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_tag_resource}
\alias{lightsail_tag_resource}
\title{Adds one or more tags to the specified Amazon Lightsail resource}
\usage{
lightsail_tag_resource(resourceName, resourceArn, tags)
}
\arguments{
\item{resourceName}{[required] The name of the resource to which you are adding tags.}

\item{resourceArn}{The Amazon Resource Name (ARN) of the resource to which you want to add
a tag.}

\item{tags}{[required] The tag key and optional value.}
}
\description{
Adds one or more tags to the specified Amazon Lightsail resource. Each
resource can have a maximum of 50 tags. Each tag consists of a key and
an optional value. Tag keys must be unique per resource. For more
information about tags, see the Lightsail Dev Guide.

The \verb{tag resource} operation supports tag-based access control via
request tags and resource tags applied to the resource identified by
\verb{resource name}. For more information, see the Lightsail Dev Guide.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceName = "string",
  resourceArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
