% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_delete_geofence}
\alias{locationservice_batch_delete_geofence}
\title{Deletes a batch of geofences from a geofence collection}
\usage{
locationservice_batch_delete_geofence(CollectionName, GeofenceIds)
}
\arguments{
\item{CollectionName}{[required] The geofence collection storing the geofences to be deleted.}

\item{GeofenceIds}{[required] The batch of geofences to be deleted.}
}
\description{
Deletes a batch of geofences from a geofence collection.
}
\details{
This action deletes the resource permanently. You can't undo this
action.
}
\section{Request syntax}{
\preformatted{svc$batch_delete_geofence(
  CollectionName = "string",
  GeofenceIds = list(
    "string"
  )
)
}
}

\keyword{internal}
