% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_evaluate_geofences}
\alias{locationservice_batch_evaluate_geofences}
\title{Used in geofence monitoring}
\usage{
locationservice_batch_evaluate_geofences(CollectionName,
  DevicePositionUpdates)
}
\arguments{
\item{CollectionName}{[required] The geofence collection used in evaluating the position of devices
against its geofences.}

\item{DevicePositionUpdates}{[required] Contains device details for each device to be evaluated against the
given geofence collection.}
}
\description{
Used in geofence monitoring. Evaluates device positions against the
position of geofences in a given geofence collection.
}
\section{Request syntax}{
\preformatted{svc$batch_evaluate_geofences(
  CollectionName = "string",
  DevicePositionUpdates = list(
    list(
      DeviceId = "string",
      Position = list(
        123.0
      ),
      SampleTime = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}

\keyword{internal}
