% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_get_device_position}
\alias{locationservice_batch_get_device_position}
\title{A batch request to retrieve device positions}
\usage{
locationservice_batch_get_device_position(DeviceIds, TrackerName)
}
\arguments{
\item{DeviceIds}{[required] Devices whose position you want to retrieve.
\itemize{
\item For example, for two devices:
\verb{device-ids=DeviceId1&amp;device-ids=DeviceId2}
}}

\item{TrackerName}{[required] The tracker resource retrieving the device position.}
}
\description{
A batch request to retrieve device positions.

The response will return the device positions from the last 24 hours.
}
\section{Request syntax}{
\preformatted{svc$batch_get_device_position(
  DeviceIds = list(
    "string"
  ),
  TrackerName = "string"
)
}
}

\keyword{internal}
