% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_put_geofence}
\alias{locationservice_batch_put_geofence}
\title{A batch request for storing geofences into a given geofence collection}
\usage{
locationservice_batch_put_geofence(CollectionName, Entries)
}
\arguments{
\item{CollectionName}{[required] The geofence collection storing the geofences.}

\item{Entries}{[required] The batch of geofences to be stored in a geofence collection.}
}
\description{
A batch request for storing geofences into a given geofence collection.
}
\section{Request syntax}{
\preformatted{svc$batch_put_geofence(
  CollectionName = "string",
  Entries = list(
    list(
      GeofenceId = "string",
      Geometry = list(
        Polygon = list(
          list(
            list(
              123.0
            )
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
