% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_create_geofence_collection}
\alias{locationservice_create_geofence_collection}
\title{Creates a geofence collection, which manages and stores geofences}
\usage{
locationservice_create_geofence_collection(CollectionName, Description,
  PricingPlan)
}
\arguments{
\item{CollectionName}{[required] A custom name for the geofence collection.

Requirements:
\itemize{
\item Contain only alphanumeric characters (A–Z, a–z, 0-9), hyphens (-),
and underscores (\\_).
\item Must be a unique geofence collection name.
\item No spaces allowed. For example, \code{ExampleGeofenceCollection}.
}}

\item{Description}{An optional description for the geofence collection.}

\item{PricingPlan}{[required] Specifies the pricing plan for your geofence collection. There's three
pricing plan options:
\itemize{
\item \code{RequestBasedUsage} — Selects the "Request-Based Usage" pricing
plan.
\item \code{MobileAssetTracking} — Selects the "Mobile Asset Tracking" pricing
plan.
\item \code{MobileAssetManagement} — Selects the "Mobile Asset Management"
pricing plan.
}

For additional details and restrictions on each pricing plan option, see
the \href{https://aws.amazon.com/location/pricing/}{Amazon Location Service pricing page}.}
}
\description{
Creates a geofence collection, which manages and stores geofences.
}
\section{Request syntax}{
\preformatted{svc$create_geofence_collection(
  CollectionName = "string",
  Description = "string",
  PricingPlan = "RequestBasedUsage"|"MobileAssetTracking"|"MobileAssetManagement"
)
}
}

\keyword{internal}
