% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_create_map}
\alias{locationservice_create_map}
\title{Creates a map resource in your AWS account, which provides map tiles of
different styles sourced from global location data providers}
\usage{
locationservice_create_map(Configuration, Description, MapName,
  PricingPlan)
}
\arguments{
\item{Configuration}{[required] Specifies the map style selected from an available data provider.}

\item{Description}{An optional description for the map resource.}

\item{MapName}{[required] The name for the map resource.

Requirements:
\itemize{
\item Must contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens
(-), and underscores (\\_).
\item Must be a unique map resource name.
\item No spaces allowed. For example, \code{ExampleMap}.
}}

\item{PricingPlan}{[required] Specifies the pricing plan for your map resource. There's three pricing
plan options:
\itemize{
\item \code{RequestBasedUsage} — Selects the "Request-Based Usage" pricing
plan.
\item \code{MobileAssetTracking} — Selects the "Mobile Asset Tracking" pricing
plan.
\item \code{MobileAssetManagement} — Selects the "Mobile Asset Management"
pricing plan.
}

For additional details and restrictions on each pricing plan option, see
the \href{https://aws.amazon.com/location/pricing/}{Amazon Location Service pricing page}.}
}
\description{
Creates a map resource in your AWS account, which provides map tiles of
different styles sourced from global location data providers.
}
\details{
By using Maps, you agree that AWS may transmit your API queries to your
selected third party provider for processing, which may be outside the
AWS region you are currently using. For more information, see the \href{https://aws.amazon.com/service-terms/}{AWS Service Terms} for Amazon
Location Service.
}
\section{Request syntax}{
\preformatted{svc$create_map(
  Configuration = list(
    Style = "string"
  ),
  Description = "string",
  MapName = "string",
  PricingPlan = "RequestBasedUsage"|"MobileAssetTracking"|"MobileAssetManagement"
)
}
}

\keyword{internal}
