% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_create_place_index}
\alias{locationservice_create_place_index}
\title{Creates a Place index resource in your AWS account, which supports
Places functions with geospatial data sourced from your chosen data
provider}
\usage{
locationservice_create_place_index(DataSource, DataSourceConfiguration,
  Description, IndexName, PricingPlan)
}
\arguments{
\item{DataSource}{[required] Specifies the data provider of geospatial data.}

\item{DataSourceConfiguration}{Specifies the data storage option for requesting Places.}

\item{Description}{The optional description for the Place index resource.}

\item{IndexName}{[required] The name of the Place index resource.

Requirements:
\itemize{
\item Contain only alphanumeric characters (A-Z, a-z, 0-9) , hyphens (-)
and underscores (\\_) ).
\item Must be a unique Place index resource name.
\item No spaces allowed. For example, \code{ExamplePlaceIndex}.
}}

\item{PricingPlan}{[required] Specifies the pricing plan for your Place index resource. There's three
pricing plan options:
\itemize{
\item \code{RequestBasedUsage} — Selects the "Request-Based Usage" pricing
plan.
\item \code{MobileAssetTracking} — Selects the "Mobile Asset Tracking" pricing
plan.
\item \code{MobileAssetManagement} — Selects the "Mobile Asset Management"
pricing plan.
}

For additional details and restrictions on each pricing plan option, see
the \href{https://aws.amazon.com/location/pricing/}{Amazon Location Service pricing page}.}
}
\description{
Creates a Place index resource in your AWS account, which supports
Places functions with geospatial data sourced from your chosen data
provider.

By using Places, you agree that AWS may transmit your API queries to
your selected third party provider for processing, which may be outside
the AWS region you are currently using.

Because of licensing limitations, you may not use HERE to store results
for locations in Japan. For more information, see the \href{https://aws.amazon.com/service-terms/}{AWS Service Terms} for Amazon Location
Service.
}
\section{Request syntax}{
\preformatted{svc$create_place_index(
  DataSource = "string",
  DataSourceConfiguration = list(
    IntendedUse = "SingleUse"|"Storage"
  ),
  Description = "string",
  IndexName = "string",
  PricingPlan = "RequestBasedUsage"|"MobileAssetTracking"|"MobileAssetManagement"
)
}
}

\keyword{internal}
