% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_create_tracker}
\alias{locationservice_create_tracker}
\title{Creates a tracker resource in your AWS account, which lets you retrieve
current and historical location of devices}
\usage{
locationservice_create_tracker(Description, PricingPlan, TrackerName)
}
\arguments{
\item{Description}{An optional description for the tracker resource.}

\item{PricingPlan}{[required] Specifies the pricing plan for your tracker resource. There's three
pricing plan options:
\itemize{
\item \code{RequestBasedUsage} — Selects the "Request-Based Usage" pricing
plan.
\item \code{MobileAssetTracking} — Selects the "Mobile Asset Tracking" pricing
plan.
\item \code{MobileAssetManagement} — Selects the "Mobile Asset Management"
pricing plan.
}

For additional details and restrictions on each pricing plan option, see
the \href{https://aws.amazon.com/location/pricing/}{Amazon Location Service pricing page}.}

\item{TrackerName}{[required] The name for the tracker resource.

Requirements:
\itemize{
\item Contain only alphanumeric characters (A-Z, a-z, 0-9) , hyphens (-)
and underscores (\\_).
\item Must be a unique tracker resource name.
\item No spaces allowed. For example, \code{ExampleTracker}.
}}
}
\description{
Creates a tracker resource in your AWS account, which lets you retrieve
current and historical location of devices.
}
\section{Request syntax}{
\preformatted{svc$create_tracker(
  Description = "string",
  PricingPlan = "RequestBasedUsage"|"MobileAssetTracking"|"MobileAssetManagement",
  TrackerName = "string"
)
}
}

\keyword{internal}
