% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_geofence_collection}
\alias{locationservice_delete_geofence_collection}
\title{Deletes a geofence collection from your AWS account}
\usage{
locationservice_delete_geofence_collection(CollectionName)
}
\arguments{
\item{CollectionName}{[required] The name of the geofence collection to be deleted.}
}
\description{
Deletes a geofence collection from your AWS account.

This action deletes the resource permanently. You can't undo this
action. If the geofence collection is the target of a tracker resource,
the devices will no longer be monitored.
}
\section{Request syntax}{
\preformatted{svc$delete_geofence_collection(
  CollectionName = "string"
)
}
}

\keyword{internal}
