% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_map}
\alias{locationservice_delete_map}
\title{Deletes a map resource from your AWS account}
\usage{
locationservice_delete_map(MapName)
}
\arguments{
\item{MapName}{[required] The name of the map resource to be deleted.}
}
\description{
Deletes a map resource from your AWS account.
}
\details{
This action deletes the resource permanently. You cannot undo this
action. If the map is being used in an application, the map may not
render.
}
\section{Request syntax}{
\preformatted{svc$delete_map(
  MapName = "string"
)
}
}

\keyword{internal}
