% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_place_index}
\alias{locationservice_delete_place_index}
\title{Deletes a Place index resource from your AWS account}
\usage{
locationservice_delete_place_index(IndexName)
}
\arguments{
\item{IndexName}{[required] The name of the Place index resource to be deleted.}
}
\description{
Deletes a Place index resource from your AWS account.
}
\details{
This action deletes the resource permanently. You cannot undo this
action.
}
\section{Request syntax}{
\preformatted{svc$delete_place_index(
  IndexName = "string"
)
}
}

\keyword{internal}
