% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_tracker}
\alias{locationservice_delete_tracker}
\title{Deletes a tracker resource from your AWS account}
\usage{
locationservice_delete_tracker(TrackerName)
}
\arguments{
\item{TrackerName}{[required] The name of the tracker resource to be deleted.}
}
\description{
Deletes a tracker resource from your AWS account.
}
\details{
This action deletes the resource permanently. You can't undo this
action. If the tracker resource is in use, you may encounter an error.
Make sure that the target resource is not a dependency for your
applications.
}
\section{Request syntax}{
\preformatted{svc$delete_tracker(
  TrackerName = "string"
)
}
}

\keyword{internal}
