% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_describe_place_index}
\alias{locationservice_describe_place_index}
\title{Retrieves the Place index resource details}
\usage{
locationservice_describe_place_index(IndexName)
}
\arguments{
\item{IndexName}{[required] The name of the Place index resource.}
}
\description{
Retrieves the Place index resource details.
}
\section{Request syntax}{
\preformatted{svc$describe_place_index(
  IndexName = "string"
)
}
}

\keyword{internal}
