% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_describe_tracker}
\alias{locationservice_describe_tracker}
\title{Retrieves the tracker resource details}
\usage{
locationservice_describe_tracker(TrackerName)
}
\arguments{
\item{TrackerName}{[required] The name of the tracker resource.}
}
\description{
Retrieves the tracker resource details.
}
\section{Request syntax}{
\preformatted{svc$describe_tracker(
  TrackerName = "string"
)
}
}

\keyword{internal}
