% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_disassociate_tracker_consumer}
\alias{locationservice_disassociate_tracker_consumer}
\title{Removes the association bewteen a tracker resource and a geofence
collection}
\usage{
locationservice_disassociate_tracker_consumer(ConsumerArn, TrackerName)
}
\arguments{
\item{ConsumerArn}{[required] The Amazon Resource Name (ARN) for the geofence collection to be
disassociated from the tracker resource. Used when you need to specify a
resource across all AWS.
\itemize{
\item Format example:
\code{arn:partition:service:region:account-id:resource-type:resource-id}
}}

\item{TrackerName}{[required] The name of the tracker resource to be dissociated from the consumer.}
}
\description{
Removes the association bewteen a tracker resource and a geofence
collection.

Once you unlink a tracker resource from a geofence collection, the
tracker positions will no longer be automatically evaluated against
geofences.
}
\section{Request syntax}{
\preformatted{svc$disassociate_tracker_consumer(
  ConsumerArn = "string",
  TrackerName = "string"
)
}
}

\keyword{internal}
