% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_device_position}
\alias{locationservice_get_device_position}
\title{Retrieves the latest device position}
\usage{
locationservice_get_device_position(DeviceId, TrackerName)
}
\arguments{
\item{DeviceId}{[required] The device whose position you want to retreieve.}

\item{TrackerName}{[required] The tracker resource receiving the position update.}
}
\description{
Retrieves the latest device position.

Limitation — Device positions are deleted after one year.
}
\section{Request syntax}{
\preformatted{svc$get_device_position(
  DeviceId = "string",
  TrackerName = "string"
)
}
}

\keyword{internal}
