% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_device_position_history}
\alias{locationservice_get_device_position_history}
\title{Retrieves the device position history from a tracker resource within a
specified range of time}
\usage{
locationservice_get_device_position_history(DeviceId, EndTimeExclusive,
  NextToken, StartTimeInclusive, TrackerName)
}
\arguments{
\item{DeviceId}{[required] The device whose position history you want to retrieve.}

\item{EndTimeExclusive}{Specify the end time for the position history in \href{https://www.iso.org/iso-8601-date-and-time-format.html}{ISO 8601} format:
\code{YYYY-MM-DDThh:mm:ss.sssZ}.
\itemize{
\item The given time for \code{EndTimeExclusive} must be after the time for
\code{StartTimeInclusive}.
}}

\item{NextToken}{The pagination token specifying which page of results to return in the
response. If no token is provided, the default page is the first page.

Default value: \code{null}}

\item{StartTimeInclusive}{Specify the start time for the position history in \href{https://www.iso.org/iso-8601-date-and-time-format.html}{ISO 8601} format:
\code{YYYY-MM-DDThh:mm:ss.sssZ}.
\itemize{
\item The given time for \code{EndTimeExclusive} must be after the time for
\code{StartTimeInclusive}.
}}

\item{TrackerName}{[required] The tracker resource receiving the request for the device position
history.}
}
\description{
Retrieves the device position history from a tracker resource within a
specified range of time.
}
\details{
Limitation — Device positions are deleted after one year.
}
\section{Request syntax}{
\preformatted{svc$get_device_position_history(
  DeviceId = "string",
  EndTimeExclusive = as.POSIXct(
    "2015-01-01"
  ),
  NextToken = "string",
  StartTimeInclusive = as.POSIXct(
    "2015-01-01"
  ),
  TrackerName = "string"
)
}
}

\keyword{internal}
