% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_map_glyphs}
\alias{locationservice_get_map_glyphs}
\title{Retrieves glyphs used to display labels on a map}
\usage{
locationservice_get_map_glyphs(FontStack, FontUnicodeRange, MapName)
}
\arguments{
\item{FontStack}{[required] A comma-separated list of fonts to load glyphs from in order of
preference.. For example, \verb{Noto Sans, Arial Unicode}.}

\item{FontUnicodeRange}{[required] A Unicode range of characters to download glyphs for. Each response will
contain 256 characters. For example, 0-255 includes all characters from
range \code{U+0000} to \verb{00FF}. Must be aligned to multiples of 256.}

\item{MapName}{[required] The map resource associated with the glyph ﬁle.}
}
\description{
Retrieves glyphs used to display labels on a map.
}
\section{Request syntax}{
\preformatted{svc$get_map_glyphs(
  FontStack = "string",
  FontUnicodeRange = "string",
  MapName = "string"
)
}
}

\keyword{internal}
