% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_map_style_descriptor}
\alias{locationservice_get_map_style_descriptor}
\title{Retrieves the map style descriptor from a map resource}
\usage{
locationservice_get_map_style_descriptor(MapName)
}
\arguments{
\item{MapName}{[required] The map resource to retrieve the style descriptor from.}
}
\description{
Retrieves the map style descriptor from a map resource.

The style descriptor contains speciﬁcations on how features render on a
map. For example, what data to display, what order to display the data
in, and the style for the data. Style descriptors follow the Mapbox
Style Specification.
}
\section{Request syntax}{
\preformatted{svc$get_map_style_descriptor(
  MapName = "string"
)
}
}

\keyword{internal}
