% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_map_tile}
\alias{locationservice_get_map_tile}
\title{Retrieves a vector data tile from the map resource}
\usage{
locationservice_get_map_tile(MapName, X, Y, Z)
}
\arguments{
\item{MapName}{[required] The map resource to retrieve the map tiles from.}

\item{X}{[required] The X axis value for the map tile.}

\item{Y}{[required] The Y axis value for the map tile.}

\item{Z}{[required] The zoom value for the map tile.}
}
\description{
Retrieves a vector data tile from the map resource. Map tiles are used
by clients to render a map. They are addressed using a grid arrangement
with an X coordinate, Y coordinate, and Z (zoom) level.
}
\details{
The origin (0, 0) is the top left of the map. Increasing the zoom level
by 1 doubles both the X and Y dimensions, so a tile containing data for
the entire world at (0/0/0) will be split into 4 tiles at zoom 1 (1/0/0,
1/0/1, 1/1/0, 1/1/1).
}
\section{Request syntax}{
\preformatted{svc$get_map_tile(
  MapName = "string",
  X = "string",
  Y = "string",
  Z = "string"
)
}
}

\keyword{internal}
