% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_put_geofence}
\alias{locationservice_put_geofence}
\title{Stores a geofence to a given geofence collection, or updates the
geometry of an existing geofence if a geofence ID is included in the
request}
\usage{
locationservice_put_geofence(CollectionName, GeofenceId, Geometry)
}
\arguments{
\item{CollectionName}{[required] The geofence collection to store the geofence in.}

\item{GeofenceId}{[required] An identifier for the geofence. For example, \code{ExampleGeofence-1}.}

\item{Geometry}{[required] Contains the polygon details to specify the position of the geofence.}
}
\description{
Stores a geofence to a given geofence collection, or updates the
geometry of an existing geofence if a geofence ID is included in the
request.
}
\section{Request syntax}{
\preformatted{svc$put_geofence(
  CollectionName = "string",
  GeofenceId = "string",
  Geometry = list(
    Polygon = list(
      list(
        list(
          123.0
        )
      )
    )
  )
)
}
}

\keyword{internal}
