% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_search_place_index_for_position}
\alias{locationservice_search_place_index_for_position}
\title{Reverse geocodes a given coordinate and returns a legible address}
\usage{
locationservice_search_place_index_for_position(IndexName, MaxResults,
  Position)
}
\arguments{
\item{IndexName}{[required] The name of the Place index resource you want to use for the search.}

\item{MaxResults}{An optional paramer. The maximum number of results returned per request.

Default value: \code{50}}

\item{Position}{[required] Specifies a coordinate for the query defined by a longitude, and
latitude.
\itemize{
\item The first position is the X coordinate, or longitude.
\item The second position is the Y coordinate, or latitude.
}

For example, \verb{position=xLongitude&amp;position=yLatitude} .}
}
\description{
Reverse geocodes a given coordinate and returns a legible address.
Allows you to search for Places or points of interest near a given
position.
}
\details{
By using Places, you agree that AWS may transmit your API queries to
your selected third party provider for processing, which may be outside
the AWS region you are currently using.

Because of licensing limitations, you may not use HERE to store results
for locations in Japan. For more information, see the \href{https://aws.amazon.com/service-terms/}{AWS Service Terms} for Amazon Location
Service.
}
\section{Request syntax}{
\preformatted{svc$search_place_index_for_position(
  IndexName = "string",
  MaxResults = 123,
  Position = list(
    123.0
  )
)
}
}

\keyword{internal}
