% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_search_place_index_for_text}
\alias{locationservice_search_place_index_for_text}
\title{Geocodes free-form text, such as an address, name, city, or region to
allow you to search for Places or points of interest}
\usage{
locationservice_search_place_index_for_text(BiasPosition, FilterBBox,
  FilterCountries, IndexName, MaxResults, Text)
}
\arguments{
\item{BiasPosition}{Searches for results closest to the given position. An optional
parameter defined by longitude, and latitude.
\itemize{
\item The first \code{bias} position is the X coordinate, or longitude.
\item The second \code{bias} position is the Y coordinate, or latitude.
}

For example, \verb{bias=xLongitude&amp;bias=yLatitude}.}

\item{FilterBBox}{Filters the results by returning only Places within the provided
bounding box. An optional parameter.

The first 2 \code{bbox} parameters describe the lower southwest corner:
\itemize{
\item The first \code{bbox} position is the X coordinate or longitude of the
lower southwest corner.
\item The second \code{bbox} position is the Y coordinate or latitude of the
lower southwest corner.
}

For example, \verb{bbox=xLongitudeSW&amp;bbox=yLatitudeSW}.

The next \code{bbox} parameters describe the upper northeast corner:
\itemize{
\item The third \code{bbox} position is the X coordinate, or longitude of the
upper northeast corner.
\item The fourth \code{bbox} position is the Y coordinate, or longitude of the
upper northeast corner.
}

For example, \verb{bbox=xLongitudeNE&amp;bbox=yLatitudeNE}}

\item{FilterCountries}{Limits the search to the given a list of countries/regions. An optional
parameter.
\itemize{
\item Use the \href{https://www.iso.org/iso-3166-country-codes.html}{ISO 3166}
3-digit country code. For example, Australia uses three upper-case
characters: \code{AUS}.
}}

\item{IndexName}{[required] The name of the Place index resource you want to use for the search.}

\item{MaxResults}{An optional parameter. The maximum number of results returned per
request.

The default: \code{50}}

\item{Text}{[required] The address, name, city, or region to be used in the search. In
free-form text format. For example, \verb{123 Any Street}.}
}
\description{
Geocodes free-form text, such as an address, name, city, or region to
allow you to search for Places or points of interest.
}
\details{
Includes the option to apply additional parameters to narrow your list
of results.

You can search for places near a given position using \code{BiasPosition}, or
filter results within a bounding box using \code{FilterBBox}. Providing both
parameters simultaneously returns an error.

By using Places, you agree that AWS may transmit your API queries to
your selected third party provider for processing, which may be outside
the AWS region you are currently using.

Also, when using HERE as your data provider, you may not (a) use HERE
Places for Asset Management, or (b) select the \code{Storage} option for the
\code{IntendedUse} parameter when requesting Places in Japan. For more
information, see the \href{https://aws.amazon.com/service-terms/}{AWS Service Terms} for Amazon Location
Service.
}
\section{Request syntax}{
\preformatted{svc$search_place_index_for_text(
  BiasPosition = list(
    123.0
  ),
  FilterBBox = list(
    123.0
  ),
  FilterCountries = list(
    "string"
  ),
  IndexName = "string",
  MaxResults = 123,
  Text = "string"
)
}
}

\keyword{internal}
