% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutforvision_operations.R
\name{lookoutforvision_create_dataset}
\alias{lookoutforvision_create_dataset}
\title{Creates a new dataset in an Amazon Lookout for Vision project}
\usage{
lookoutforvision_create_dataset(ProjectName, DatasetType, DatasetSource,
  ClientToken)
}
\arguments{
\item{ProjectName}{[required] The name of the project in which you want to create a dataset.}

\item{DatasetType}{[required] The type of the dataset. Specify \code{train} for a training dataset. Specify
\code{test} for a test dataset.}

\item{DatasetSource}{The location of the manifest file that Amazon Lookout for Vision uses to
create the dataset.

If you don't specify \code{DatasetSource}, an empty dataset is created and
the operation synchronously returns. Later, you can add JSON Lines by
calling UpdateDatasetEntries.

If you specify a value for \code{DataSource}, the manifest at the S3 location
is validated and used to create the dataset. The call to \code{CreateDataset}
is asynchronous and might take a while to complete. To find out the
current status, Check the value of \code{Status} returned in a call to
DescribeDataset.}

\item{ClientToken}{ClientToken is an idempotency token that ensures a call to
\code{CreateDataset} completes only once. You choose the value to pass. For
example, An issue, such as an network outage, might prevent you from
getting a response from \code{CreateDataset}. In this case, safely retry your
call to \code{CreateDataset} by using the same \code{ClientToken} parameter value.
An error occurs if the other input parameters are not the same as in the
first request. Using a different value for \code{ClientToken} is considered a
new call to \code{CreateDataset}. An idempotency token is active for 8 hours.}
}
\description{
Creates a new dataset in an Amazon Lookout for Vision project.
\code{CreateDataset} can create a training or a test dataset from a valid
dataset source (\code{DatasetSource}).
}
\details{
If you want a single dataset project, specify \code{train} for the value of
\code{DatasetType}.

To have a project with separate training and test datasets, call
\code{CreateDataset} twice. On the first call, specify \code{train} for the value
of \code{DatasetType}. On the second call, specify \code{test} for the value of
\code{DatasetType}. of dataset with
}
\section{Request syntax}{
\preformatted{svc$create_dataset(
  ProjectName = "string",
  DatasetType = "string",
  DatasetSource = list(
    GroundTruthManifest = list(
      S3Object = list(
        Bucket = "string",
        Key = "string",
        VersionId = "string"
      )
    )
  ),
  ClientToken = "string"
)
}
}

\keyword{internal}
