% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutforvision_operations.R
\name{lookoutforvision_create_project}
\alias{lookoutforvision_create_project}
\title{Creates an empty Amazon Lookout for Vision project}
\usage{
lookoutforvision_create_project(ProjectName, ClientToken)
}
\arguments{
\item{ProjectName}{[required] S nsme for the project.}

\item{ClientToken}{ClientToken is an idempotency token that ensures a call to
\code{CreateProject} completes only once. You choose the value to pass. For
example, An issue, such as an network outage, might prevent you from
getting a response from \code{CreateProject}. In this case, safely retry your
call to \code{CreateProject} by using the same \code{ClientToken} parameter value.
An error occurs if the other input parameters are not the same as in the
first request. Using a different value for \code{ClientToken} is considered a
new call to \code{CreateProject}. An idempotency token is active for 8 hours.}
}
\description{
Creates an empty Amazon Lookout for Vision project. After you create the
project, add a dataset by calling CreateDataset.
}
\section{Request syntax}{
\preformatted{svc$create_project(
  ProjectName = "string",
  ClientToken = "string"
)
}
}

\keyword{internal}
