% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutforvision_operations.R
\name{lookoutforvision_delete_dataset}
\alias{lookoutforvision_delete_dataset}
\title{Deletes an existing Amazon Lookout for Vision dataset}
\usage{
lookoutforvision_delete_dataset(ProjectName, DatasetType, ClientToken)
}
\arguments{
\item{ProjectName}{[required] The name of the project that contains the dataset that you want to
delete.}

\item{DatasetType}{[required] The type of the dataset to delete. Specify \code{train} to delete the
training dataset. Specify \code{test} to delete the test dataset. To delete
the dataset in a single dataset project, specify \code{train}.}

\item{ClientToken}{ClientToken is an idempotency token that ensures a call to
\code{DeleteDataset} completes only once. You choose the value to pass. For
example, An issue, such as an network outage, might prevent you from
getting a response from \code{DeleteDataset}. In this case, safely retry your
call to \code{DeleteDataset} by using the same \code{ClientToken} parameter value.
An error occurs if the other input parameters are not the same as in the
first request. Using a different value for \code{ClientToken} is considered a
new call to \code{DeleteDataset}. An idempotency token is active for 8 hours.}
}
\description{
Deletes an existing Amazon Lookout for Vision \code{dataset}.
}
\details{
If your the project has a single dataset, you must create a new dataset
before you can create a model.

If you project has a training dataset and a test dataset consider the
following.
\itemize{
\item If you delete the test dataset, your project reverts to a single
dataset project. If you then train the model, Amazon Lookout for
Vision internally splits the remaining dataset into a training and
test dataset.
\item If you delete the training dataset, you must create a training
dataset before you can create a model.
}

It might take a while to delete the dataset. To check the current
status, check the \code{Status} field in the response from a call to
DescribeDataset.
}
\section{Request syntax}{
\preformatted{svc$delete_dataset(
  ProjectName = "string",
  DatasetType = "string",
  ClientToken = "string"
)
}
}

\keyword{internal}
