% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutforvision_operations.R
\name{lookoutforvision_delete_project}
\alias{lookoutforvision_delete_project}
\title{Deletes an Amazon Lookout for Vision project}
\usage{
lookoutforvision_delete_project(ProjectName, ClientToken)
}
\arguments{
\item{ProjectName}{[required] The name of the project to delete.}

\item{ClientToken}{ClientToken is an idempotency token that ensures a call to
\code{DeleteProject} completes only once. You choose the value to pass. For
example, An issue, such as an network outage, might prevent you from
getting a response from \code{DeleteProject}. In this case, safely retry your
call to \code{DeleteProject} by using the same \code{ClientToken} parameter value.
An error occurs if the other input parameters are not the same as in the
first request. Using a different value for \code{ClientToken} is considered a
new call to \code{DeleteProject}. An idempotency token is active for 8 hours.}
}
\description{
Deletes an Amazon Lookout for Vision project.

To delete a project, you must first delete each version of the model
associated with the project. To delete a model use the DeleteModel
operation.

The training and test datasets are deleted automatically for you. The
images referenced by the training and test datasets aren't deleted.
}
\section{Request syntax}{
\preformatted{svc$delete_project(
  ProjectName = "string",
  ClientToken = "string"
)
}
}

\keyword{internal}
