% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutforvision_operations.R
\name{lookoutforvision_describe_dataset}
\alias{lookoutforvision_describe_dataset}
\title{Describe an Amazon Lookout for Vision dataset}
\usage{
lookoutforvision_describe_dataset(ProjectName, DatasetType)
}
\arguments{
\item{ProjectName}{[required] The name of the project that contains the dataset that you want to
describe.}

\item{DatasetType}{[required] The type of the dataset to describe. Specify \code{train} to describe the
training dataset. Specify \code{test} to describe the test dataset. If you
have a single dataset project, specify \code{train}}
}
\description{
Describe an Amazon Lookout for Vision dataset.
}
\section{Request syntax}{
\preformatted{svc$describe_dataset(
  ProjectName = "string",
  DatasetType = "string"
)
}
}

\keyword{internal}
