% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutforvision_operations.R
\name{lookoutforvision_describe_model}
\alias{lookoutforvision_describe_model}
\title{Describes a version of an Amazon Lookout for Vision model}
\usage{
lookoutforvision_describe_model(ProjectName, ModelVersion)
}
\arguments{
\item{ProjectName}{[required] The project that contains the version of a model that you want to
describe.}

\item{ModelVersion}{[required] The version of the model that you want to describe.}
}
\description{
Describes a version of an Amazon Lookout for Vision model.
}
\section{Request syntax}{
\preformatted{svc$describe_model(
  ProjectName = "string",
  ModelVersion = "string"
)
}
}

\keyword{internal}
