% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutforvision_operations.R
\name{lookoutforvision_describe_project}
\alias{lookoutforvision_describe_project}
\title{Describes an Amazon Lookout for Vision project}
\usage{
lookoutforvision_describe_project(ProjectName)
}
\arguments{
\item{ProjectName}{[required] The name of the project that you want to describe.}
}
\description{
Describes an Amazon Lookout for Vision project.
}
\section{Request syntax}{
\preformatted{svc$describe_project(
  ProjectName = "string"
)
}
}

\keyword{internal}
