% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutforvision_operations.R
\name{lookoutforvision_list_dataset_entries}
\alias{lookoutforvision_list_dataset_entries}
\title{Lists the JSON Lines within a dataset}
\usage{
lookoutforvision_list_dataset_entries(ProjectName, DatasetType, Labeled,
  AnomalyClass, BeforeCreationDate, AfterCreationDate, NextToken,
  MaxResults, SourceRefContains)
}
\arguments{
\item{ProjectName}{[required] The name of the project that contains the dataset that you want to list.}

\item{DatasetType}{[required] The type of the dataset that you want to list. Specify \code{train} to list
the training dataset. Specify \code{test} to list the test dataset. If you
have a single dataset project, specify \code{train}.}

\item{Labeled}{Specify \code{true} to include labeled entries, otherwise specify \code{false}. If
you don't specify a value, Lookout for Vision returns all entries.}

\item{AnomalyClass}{Specify \code{normal} to include only normal images. Specify \code{anomaly} to
only include anomalous entries. If you don't specify a value, Amazon
Lookout for Vision returns normal and anomalous images.}

\item{BeforeCreationDate}{Only includes entries before the specified date in the response. For
example, \verb{2020-06-23T00:00:00}.}

\item{AfterCreationDate}{Only includes entries after the specified date in the response. For
example, \verb{2020-06-23T00:00:00}.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Lookout for Vision returns a pagination token in the
response. You can use this pagination token to retrieve the next set of
dataset entries.}

\item{MaxResults}{The maximum number of results to return per paginated call. The largest
value you can specify is 100. If you specify a value greater than 100, a
ValidationException error occurs. The default value is 100.}

\item{SourceRefContains}{Perform a "contains" search on the values of the \code{source-ref} key within
the dataset. For example a value of "IMG\\_17" returns all JSON Lines
where the \code{source-ref} key value matches \strong{IMG\\_17}.}
}
\description{
Lists the JSON Lines within a dataset. An Amazon Lookout for Vision JSON
Line contains the anomaly information for a single image, including the
image location and the assigned label.
}
\section{Request syntax}{
\preformatted{svc$list_dataset_entries(
  ProjectName = "string",
  DatasetType = "string",
  Labeled = TRUE|FALSE,
  AnomalyClass = "string",
  BeforeCreationDate = as.POSIXct(
    "2015-01-01"
  ),
  AfterCreationDate = as.POSIXct(
    "2015-01-01"
  ),
  NextToken = "string",
  MaxResults = 123,
  SourceRefContains = "string"
)
}
}

\keyword{internal}
