% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutforvision_operations.R
\name{lookoutforvision_list_models}
\alias{lookoutforvision_list_models}
\title{Lists the versions of a model in an Amazon Lookout for Vision project}
\usage{
lookoutforvision_list_models(ProjectName, NextToken, MaxResults)
}
\arguments{
\item{ProjectName}{[required] The name of the project that contains the model versions that you want
to list.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Lookout for Vision returns a pagination token in the
response. You can use this pagination token to retrieve the next set of
models.}

\item{MaxResults}{The maximum number of results to return per paginated call. The largest
value you can specify is 100. If you specify a value greater than 100, a
ValidationException error occurs. The default value is 100.}
}
\description{
Lists the versions of a model in an Amazon Lookout for Vision project.
}
\section{Request syntax}{
\preformatted{svc$list_models(
  ProjectName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
