% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_get_batch_prediction}
\alias{machinelearning_get_batch_prediction}
\title{Returns a BatchPrediction that includes detailed metadata, status, and
data file information for a Batch Prediction request}
\usage{
machinelearning_get_batch_prediction(BatchPredictionId)
}
\arguments{
\item{BatchPredictionId}{[required] An ID assigned to the \code{BatchPrediction} at creation.}
}
\description{
Returns a \code{BatchPrediction} that includes detailed metadata, status, and
data file information for a \verb{Batch Prediction} request.
}
\section{Request syntax}{
\preformatted{svc$get_batch_prediction(
  BatchPredictionId = "string"
)
}
}

\keyword{internal}
