% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_predict}
\alias{machinelearning_predict}
\title{Generates a prediction for the observation using the specified ML Model}
\usage{
machinelearning_predict(MLModelId, Record, PredictEndpoint)
}
\arguments{
\item{MLModelId}{[required] A unique identifier of the \code{MLModel}.}

\item{Record}{[required]}

\item{PredictEndpoint}{[required]}
}
\description{
Generates a prediction for the observation using the specified
\verb{ML Model}.
}
\details{
Note

Not all response parameters will be populated. Whether a response
parameter is populated depends on the type of model requested.
}
\section{Request syntax}{
\preformatted{svc$predict(
  MLModelId = "string",
  Record = list(
    "string"
  ),
  PredictEndpoint = "string"
)
}
}

\keyword{internal}
