% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_batch_get_custom_data_identifiers}
\alias{macie2_batch_get_custom_data_identifiers}
\title{Retrieves information about one or more custom data identifiers}
\usage{
macie2_batch_get_custom_data_identifiers(ids)
}
\arguments{
\item{ids}{An array of strings that lists the unique identifiers for the custom
data identifiers to retrieve information about.}
}
\description{
Retrieves information about one or more custom data identifiers.
}
\section{Request syntax}{
\preformatted{svc$batch_get_custom_data_identifiers(
  ids = list(
    "string"
  )
)
}
}

\keyword{internal}
