% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_create_classification_job}
\alias{macie2_create_classification_job}
\title{Creates and defines the settings for a classification job}
\usage{
macie2_create_classification_job(clientToken, customDataIdentifierIds,
  description, initialRun, jobType, name, s3JobDefinition,
  samplingPercentage, scheduleFrequency, tags)
}
\arguments{
\item{clientToken}{[required] A unique, case-sensitive token that you provide to ensure the
idempotency of the request.}

\item{customDataIdentifierIds}{The custom data identifiers to use for data analysis and classification.}

\item{description}{A custom description of the job. The description can contain as many as
200 characters.}

\item{initialRun}{Specifies whether to analyze all existing, eligible objects immediately
after the job is created.}

\item{jobType}{[required] The schedule for running the job. Valid values are:
\itemize{
\item ONE\\_TIME - Run the job only once. If you specify this value, don't
specify a value for the scheduleFrequency property.
\item SCHEDULED - Run the job on a daily, weekly, or monthly basis. If you
specify this value, use the scheduleFrequency property to define the
recurrence pattern for the job.
}}

\item{name}{[required] A custom name for the job. The name can contain as many as 500
characters.}

\item{s3JobDefinition}{[required] The S3 buckets that contain the objects to analyze, and the scope of
that analysis.}

\item{samplingPercentage}{The sampling depth, as a percentage, to apply when processing objects.
This value determines the percentage of eligible objects that the job
analyzes. If this value is less than 100, Amazon Macie selects the
objects to analyze at random, up to the specified percentage, and
analyzes all the data in those objects.}

\item{scheduleFrequency}{The recurrence pattern for running the job. To run the job only once,
don't specify a value for this property and set the value for the
jobType property to ONE\\_TIME.}

\item{tags}{A map of key-value pairs that specifies the tags to associate with the
job.

A job can have a maximum of 50 tags. Each tag consists of a tag key and
an associated tag value. The maximum length of a tag key is 128
characters. The maximum length of a tag value is 256 characters.}
}
\description{
<p>Creates and defines the settings for a classification job.</p>
}
\section{Request syntax}{
\preformatted{svc$create_classification_job(
  clientToken = "string",
  customDataIdentifierIds = list(
    "string"
  ),
  description = "string",
  initialRun = TRUE|FALSE,
  jobType = "ONE_TIME"|"SCHEDULED",
  name = "string",
  s3JobDefinition = list(
    bucketDefinitions = list(
      list(
        accountId = "string",
        buckets = list(
          "string"
        )
      )
    ),
    scoping = list(
      excludes = list(
        and = list(
          list(
            simpleScopeTerm = list(
              comparator = "EQ"|"GT"|"GTE"|"LT"|"LTE"|"NE"|"CONTAINS",
              key = "BUCKET_CREATION_DATE"|"OBJECT_EXTENSION"|"OBJECT_LAST_MODIFIED_DATE"|"OBJECT_SIZE"|"TAG",
              values = list(
                "string"
              )
            ),
            tagScopeTerm = list(
              comparator = "EQ"|"GT"|"GTE"|"LT"|"LTE"|"NE"|"CONTAINS",
              key = "string",
              tagValues = list(
                list(
                  key = "string",
                  value = "string"
                )
              ),
              target = "S3_OBJECT"
            )
          )
        )
      ),
      includes = list(
        and = list(
          list(
            simpleScopeTerm = list(
              comparator = "EQ"|"GT"|"GTE"|"LT"|"LTE"|"NE"|"CONTAINS",
              key = "BUCKET_CREATION_DATE"|"OBJECT_EXTENSION"|"OBJECT_LAST_MODIFIED_DATE"|"OBJECT_SIZE"|"TAG",
              values = list(
                "string"
              )
            ),
            tagScopeTerm = list(
              comparator = "EQ"|"GT"|"GTE"|"LT"|"LTE"|"NE"|"CONTAINS",
              key = "string",
              tagValues = list(
                list(
                  key = "string",
                  value = "string"
                )
              ),
              target = "S3_OBJECT"
            )
          )
        )
      )
    )
  ),
  samplingPercentage = 123,
  scheduleFrequency = list(
    dailySchedule = list(),
    monthlySchedule = list(
      dayOfMonth = 123
    ),
    weeklySchedule = list(
      dayOfWeek = "SUNDAY"|"MONDAY"|"TUESDAY"|"WEDNESDAY"|"THURSDAY"|"FRIDAY"|"SATURDAY"
    )
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
