% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_create_custom_data_identifier}
\alias{macie2_create_custom_data_identifier}
\title{Creates and defines the criteria and other settings for a custom data
identifier}
\usage{
macie2_create_custom_data_identifier(clientToken, description,
  ignoreWords, keywords, maximumMatchDistance, name, regex, tags)
}
\arguments{
\item{clientToken}{A unique, case-sensitive token that you provide to ensure the
idempotency of the request.}

\item{description}{A custom description of the custom data identifier. The description can
contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the
description of a custom data identifier. Other users of your account
might be able to see the identifier's description, depending on the
actions that they're allowed to perform in Amazon Macie.}

\item{ignoreWords}{An array that lists specific character sequences (ignore words) to
exclude from the results. If the text matched by the regular expression
is the same as any string in this array, Amazon Macie ignores it. The
array can contain as many as 10 ignore words. Each ignore word can
contain 4 - 90 characters.}

\item{keywords}{An array that lists specific character sequences (keywords), one of
which must be within proximity (maximumMatchDistance) of the regular
expression to match. The array can contain as many as 50 keywords. Each
keyword can contain 4 - 90 characters.}

\item{maximumMatchDistance}{The maximum number of characters that can exist between text that
matches the regex pattern and the character sequences specified by the
keywords array. Macie includes or excludes a result based on the
proximity of a keyword to text that matches the regex pattern. The
distance can be 1 - 300 characters. The default value is 50.}

\item{name}{A custom name for the custom data identifier. The name can contain as
many as 128 characters.

We strongly recommend that you avoid including any sensitive data in the
name of a custom data identifier. Other users of your account might be
able to see the identifier's name, depending on the actions that they're
allowed to perform in Amazon Macie.}

\item{regex}{The regular expression (\emph{regex}) that defines the pattern to match. The
expression can contain as many as 512 characters.}

\item{tags}{A map of key-value pairs that specifies the tags to associate with the
custom data identifier.

A custom data identifier can have a maximum of 50 tags. Each tag
consists of a required tag key and an associated tag value. The maximum
length of a tag key is 128 characters. The maximum length of a tag value
is 256 characters.}
}
\description{
Creates and defines the criteria and other settings for a custom data
identifier.
}
\section{Request syntax}{
\preformatted{svc$create_custom_data_identifier(
  clientToken = "string",
  description = "string",
  ignoreWords = list(
    "string"
  ),
  keywords = list(
    "string"
  ),
  maximumMatchDistance = 123,
  name = "string",
  regex = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
