% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_create_findings_filter}
\alias{macie2_create_findings_filter}
\title{Creates and defines the criteria and other settings for a findings
filter}
\usage{
macie2_create_findings_filter(action, clientToken, description,
  findingCriteria, name, position, tags)
}
\arguments{
\item{action}{[required] The action to perform on findings that meet the filter criteria
(findingCriteria). Valid values are: ARCHIVE, suppress (automatically
archive) the findings; and, NOOP, don\'t perform any action on the
findings.}

\item{clientToken}{A unique, case-sensitive token that you provide to ensure the
idempotency of the request.}

\item{description}{A custom description of the filter. The description can contain as many
as 512 characters.

We strongly recommend that you avoid including any sensitive data in the
description of a filter. Other users of your account might be able to
see the filter\'s description, depending on the actions that they\'re
allowed to perform in Amazon Macie.}

\item{findingCriteria}{[required] The criteria to use to filter findings.}

\item{name}{[required] A custom name for the filter. The name must contain at least 3
characters and can contain as many as 64 characters.

We strongly recommend that you avoid including any sensitive data in the
name of a filter. Other users of your account might be able to see the
filter\'s name, depending on the actions that they\'re allowed to
perform in Amazon Macie.}

\item{position}{The position of the filter in the list of saved filters on the Amazon
Macie console. This value also determines the order in which the filter
is applied to findings, relative to other filters that are also applied
to the findings.}

\item{tags}{A map of key-value pairs that specifies the tags to associate with the
filter.

A findings filter can have a maximum of 50 tags. Each tag consists of a
required tag key and an associated tag value. The maximum length of a
tag key is 128 characters. The maximum length of a tag value is 256
characters.}
}
\description{
Creates and defines the criteria and other settings for a findings
filter.
}
\section{Request syntax}{
\preformatted{svc$create_findings_filter(
  action = "ARCHIVE"|"NOOP",
  clientToken = "string",
  description = "string",
  findingCriteria = list(
    criterion = list(
      list(
        eq = list(
          "string"
        ),
        gt = 123,
        gte = 123,
        lt = 123,
        lte = 123,
        neq = list(
          "string"
        )
      )
    )
  ),
  name = "string",
  position = 123,
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
