% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_create_invitations}
\alias{macie2_create_invitations}
\title{Sends an Amazon Macie membership invitation to one or more accounts}
\usage{
macie2_create_invitations(accountIds, disableEmailNotification, message)
}
\arguments{
\item{accountIds}{[required] An array that lists AWS account IDs, one for each account to send the
invitation to.}

\item{disableEmailNotification}{Specifies whether to send an email notification to the root user of each
account that the invitation will be sent to. This notification is in
addition to an alert that the root user receives in AWS Personal Health
Dashboard. To send an email notification to the root user of each
account, set this value to true.}

\item{message}{A custom message to include in the invitation. Amazon Macie adds this
message to the standard content that it sends for an invitation.}
}
\description{
<p>Sends an Amazon Macie membership invitation to one or more accounts.</p>
}
\section{Request syntax}{
\preformatted{svc$create_invitations(
  accountIds = list(
    "string"
  ),
  disableEmailNotification = TRUE|FALSE,
  message = "string"
)
}
}

\keyword{internal}
