% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_create_member}
\alias{macie2_create_member}
\title{Associates an account with an Amazon Macie master account}
\usage{
macie2_create_member(account, tags)
}
\arguments{
\item{account}{[required] The details for the account to associate with the master account.}

\item{tags}{A map of key-value pairs that specifies the tags to associate with the
account in Amazon Macie.

An account can have a maximum of 50 tags. Each tag consists of a
required tag key and an associated tag value. The maximum length of a
tag key is 128 characters. The maximum length of a tag value is 256
characters.}
}
\description{
<p>Associates an account with an Amazon Macie master account.</p>
}
\section{Request syntax}{
\preformatted{svc$create_member(
  account = list(
    accountId = "string",
    email = "string"
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
