% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_create_sample_findings}
\alias{macie2_create_sample_findings}
\title{Creates sample findings}
\usage{
macie2_create_sample_findings(findingTypes)
}
\arguments{
\item{findingTypes}{An array that lists one or more types of findings to include in the set
of sample findings. Currently, the only supported value is
Policy:IAMUser/S3BucketEncryptionDisabled.}
}
\description{
<p>Creates sample findings.</p>
}
\section{Request syntax}{
\preformatted{svc$create_sample_findings(
  findingTypes = list(
    "SensitiveData:S3Object/Multiple"|"SensitiveData:S3Object/Financial"|"SensitiveData:S3Object/Personal"|"SensitiveData:S3Object/Credentials"|"SensitiveData:S3Object/CustomIdentifier"|"Policy:IAMUser/S3BucketPublic"|"Policy:IAMUser/S3BucketSharedExternally"|"Policy:IAMUser/S3BucketReplicatedExternally"|"Policy:IAMUser/S3BucketEncryptionDisabled"|"Policy:IAMUser/S3BlockPublicAccessDisabled"
  )
)
}
}

\keyword{internal}
