% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_disable_macie}
\alias{macie2_disable_macie}
\title{Disables an Amazon Macie account and deletes Macie resources for the
account}
\usage{
macie2_disable_macie()
}
\description{
Disables an Amazon Macie account and deletes Macie resources for the
account.
}
\section{Request syntax}{
\preformatted{svc$disable_macie()
}
}

\keyword{internal}
