% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_disable_organization_admin_account}
\alias{macie2_disable_organization_admin_account}
\title{Disables an account as the delegated Amazon Macie administrator account
for an AWS organization}
\usage{
macie2_disable_organization_admin_account(adminAccountId)
}
\arguments{
\item{adminAccountId}{[required] The AWS account ID of the delegated administrator account.}
}
\description{
Disables an account as the delegated Amazon Macie administrator account
for an AWS organization.
}
\section{Request syntax}{
\preformatted{svc$disable_organization_admin_account(
  adminAccountId = "string"
)
}
}

\keyword{internal}
