% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_enable_macie}
\alias{macie2_enable_macie}
\title{Enables Amazon Macie and specifies the configuration settings for a
Macie account}
\usage{
macie2_enable_macie(clientToken, findingPublishingFrequency, status)
}
\arguments{
\item{clientToken}{A unique, case-sensitive token that you provide to ensure the
idempotency of the request.}

\item{findingPublishingFrequency}{Specifies how often to publish updates to policy findings for the account. This includes publishing updates to AWS Security Hub and Amazon EventBridge (formerly called Amazon CloudWatch Events).}

\item{status}{Specifies the status for the account. To enable Amazon Macie and start
all Amazon Macie activities for the account, set this value to ENABLED.}
}
\description{
Enables Amazon Macie and specifies the configuration settings for a
Macie account.
}
\section{Request syntax}{
\preformatted{svc$enable_macie(
  clientToken = "string",
  findingPublishingFrequency = "FIFTEEN_MINUTES"|"ONE_HOUR"|"SIX_HOURS",
  status = "PAUSED"|"ENABLED"
)
}
}

\keyword{internal}
