% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_finding_statistics}
\alias{macie2_get_finding_statistics}
\title{Retrieves (queries) aggregated statistical data about findings}
\usage{
macie2_get_finding_statistics(findingCriteria, groupBy, size,
  sortCriteria)
}
\arguments{
\item{findingCriteria}{The criteria to use to filter the query results.}

\item{groupBy}{[required] The finding property to use to group the query results. Valid values
are:
\itemize{
\item classificationDetails.jobId - The unique identifier for the
classification job that produced the finding.
\item resourcesAffected.s3Bucket.name - The name of the S3 bucket that the
finding applies to.
\item severity.description - The severity of the finding, such as High or
Medium.
\item type - The type of finding, such as Policy:IAMUser/S3BucketPublic
and SensitiveData:S3Object/Personal.
}}

\item{size}{The maximum number of items to include in each page of the response.}

\item{sortCriteria}{The criteria to use to sort the query results.}
}
\description{
<p>Retrieves (queries) aggregated statistical data about findings.</p>
}
\section{Request syntax}{
\preformatted{svc$get_finding_statistics(
  findingCriteria = list(
    criterion = list(
      list(
        eq = list(
          "string"
        ),
        gt = 123,
        gte = 123,
        lt = 123,
        lte = 123,
        neq = list(
          "string"
        )
      )
    )
  ),
  groupBy = "resourcesAffected.s3Bucket.name"|"type"|"classificationDetails.jobId"|"severity.description",
  size = 123,
  sortCriteria = list(
    attributeName = "groupKey"|"count",
    orderBy = "ASC"|"DESC"
  )
)
}
}

\keyword{internal}
