% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_findings}
\alias{macie2_get_findings}
\title{Retrieves information about one or more findings}
\usage{
macie2_get_findings(findingIds, sortCriteria)
}
\arguments{
\item{findingIds}{[required] An array of strings that lists the unique identifiers for the findings
to retrieve information about.}

\item{sortCriteria}{The criteria for sorting the results of the request.}
}
\description{
Retrieves information about one or more findings.
}
\section{Request syntax}{
\preformatted{svc$get_findings(
  findingIds = list(
    "string"
  ),
  sortCriteria = list(
    attributeName = "string",
    orderBy = "ASC"|"DESC"
  )
)
}
}

\keyword{internal}
