% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_invitations_count}
\alias{macie2_get_invitations_count}
\title{Retrieves the count of Amazon Macie membership invitations that were
received by an account}
\usage{
macie2_get_invitations_count()
}
\description{
Retrieves the count of Amazon Macie membership invitations that were
received by an account.
}
\section{Request syntax}{
\preformatted{svc$get_invitations_count()
}
}

\keyword{internal}
