% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_macie_session}
\alias{macie2_get_macie_session}
\title{Retrieves information about the current status and configuration
settings for an Amazon Macie account}
\usage{
macie2_get_macie_session()
}
\description{
Retrieves information about the current status and configuration
settings for an Amazon Macie account.
}
\section{Request syntax}{
\preformatted{svc$get_macie_session()
}
}

\keyword{internal}
