% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_member}
\alias{macie2_get_member}
\title{Retrieves information about a member account that's associated with an
Amazon Macie master account}
\usage{
macie2_get_member(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource or account that the
request applies to.}
}
\description{
Retrieves information about a member account that\'s associated with an
Amazon Macie master account.
}
\section{Request syntax}{
\preformatted{svc$get_member(
  id = "string"
)
}
}

\keyword{internal}
