% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_usage_totals}
\alias{macie2_get_usage_totals}
\title{Retrieves (queries) aggregated usage data for an account}
\usage{
macie2_get_usage_totals()
}
\description{
Retrieves (queries) aggregated usage data for an account.
}
\section{Request syntax}{
\preformatted{svc$get_usage_totals()
}
}

\keyword{internal}
