% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_classification_jobs}
\alias{macie2_list_classification_jobs}
\title{Retrieves a subset of information about one or more classification jobs}
\usage{
macie2_list_classification_jobs(filterCriteria, maxResults, nextToken,
  sortCriteria)
}
\arguments{
\item{filterCriteria}{The criteria to use to filter the results.}

\item{maxResults}{The maximum number of items to include in each page of the response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}

\item{sortCriteria}{The criteria to use to sort the results.}
}
\description{
Retrieves a subset of information about one or more classification jobs.
}
\section{Request syntax}{
\preformatted{svc$list_classification_jobs(
  filterCriteria = list(
    excludes = list(
      list(
        comparator = "EQ"|"GT"|"GTE"|"LT"|"LTE"|"NE"|"CONTAINS",
        key = "jobType"|"jobStatus"|"createdAt"|"name",
        values = list(
          "string"
        )
      )
    ),
    includes = list(
      list(
        comparator = "EQ"|"GT"|"GTE"|"LT"|"LTE"|"NE"|"CONTAINS",
        key = "jobType"|"jobStatus"|"createdAt"|"name",
        values = list(
          "string"
        )
      )
    )
  ),
  maxResults = 123,
  nextToken = "string",
  sortCriteria = list(
    attributeName = "createdAt"|"jobStatus"|"name"|"jobType",
    orderBy = "ASC"|"DESC"
  )
)
}
}

\keyword{internal}
