% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_invitations}
\alias{macie2_list_invitations}
\title{Retrieves information about all the Amazon Macie membership invitations
that were received by an account}
\usage{
macie2_list_invitations(maxResults, nextToken)
}
\arguments{
\item{maxResults}{The maximum number of items to include in each page of a paginated
response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about all the Amazon Macie membership invitations
that were received by an account.
}
\section{Request syntax}{
\preformatted{svc$list_invitations(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
