% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_organization_admin_accounts}
\alias{macie2_list_organization_admin_accounts}
\title{Retrieves information about the delegated Amazon Macie administrator
account for an AWS organization}
\usage{
macie2_list_organization_admin_accounts(maxResults, nextToken)
}
\arguments{
\item{maxResults}{The maximum number of items to include in each page of a paginated
response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about the delegated Amazon Macie administrator
account for an AWS organization.
}
\section{Request syntax}{
\preformatted{svc$list_organization_admin_accounts(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
