% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_tag_resource}
\alias{macie2_tag_resource}
\title{Adds or updates one or more tags (keys and values) that are associated
with a classification job, custom data identifier, findings filter, or
member account}
\usage{
macie2_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the classification job, custom data
identifier, findings filter, or member account.}

\item{tags}{[required] A map of key-value pairs that specifies the tags to associate with the
resource.

A resource can have a maximum of 50 tags. Each tag consists of a tag key
and an associated tag value. The maximum length of a tag key is 128
characters. The maximum length of a tag value is 256 characters.}
}
\description{
Adds or updates one or more tags (keys and values) that are associated
with a classification job, custom data identifier, findings filter, or
member account.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
