% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_classification_job}
\alias{macie2_update_classification_job}
\title{Cancels a classification job}
\usage{
macie2_update_classification_job(jobId, jobStatus)
}
\arguments{
\item{jobId}{[required] The unique identifier for the classification job.}

\item{jobStatus}{[required] The status to change the job's status to. The only supported value is
CANCELLED, which cancels the job completely.}
}
\description{
Cancels a classification job.
}
\section{Request syntax}{
\preformatted{svc$update_classification_job(
  jobId = "string",
  jobStatus = "RUNNING"|"PAUSED"|"CANCELLED"|"COMPLETE"|"IDLE"
)
}
}

\keyword{internal}
