% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_service.R
\name{managedblockchain}
\alias{managedblockchain}
\title{Amazon Managed Blockchain}
\usage{
managedblockchain(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Managed Blockchain is a fully managed service for creating and
managing blockchain networks using open source frameworks. Blockchain
allows you to build applications where multiple parties can securely and
transparently run transactions and share data without the need for a
trusted, central authority. Currently, Managed Blockchain supports the
Hyperledger Fabric open source framework.
}
\section{Service syntax}{
\preformatted{svc <- managedblockchain(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=managedblockchain_create_member]{create_member} \tab Creates a member within a Managed Blockchain network \cr
\link[=managedblockchain_create_network]{create_network} \tab Creates a new blockchain network using Amazon Managed Blockchain \cr
\link[=managedblockchain_create_node]{create_node} \tab Creates a peer node in a member \cr
\link[=managedblockchain_create_proposal]{create_proposal} \tab Creates a proposal for a change to the network that other members of the network can vote on, for example, a proposal to add a new member to the network\cr
\link[=managedblockchain_delete_member]{delete_member} \tab Deletes a member \cr
\link[=managedblockchain_delete_node]{delete_node} \tab Deletes a peer node from a member that your AWS account owns \cr
\link[=managedblockchain_get_member]{get_member} \tab Returns detailed information about a member \cr
\link[=managedblockchain_get_network]{get_network} \tab Returns detailed information about a network \cr
\link[=managedblockchain_get_node]{get_node} \tab Returns detailed information about a peer node \cr
\link[=managedblockchain_get_proposal]{get_proposal} \tab Returns detailed information about a proposal \cr
\link[=managedblockchain_list_invitations]{list_invitations} \tab Returns a listing of all invitations made on the specified network \cr
\link[=managedblockchain_list_members]{list_members} \tab Returns a listing of the members in a network and properties of their configurations \cr
\link[=managedblockchain_list_networks]{list_networks} \tab Returns information about the networks in which the current AWS account has members \cr
\link[=managedblockchain_list_nodes]{list_nodes} \tab Returns information about the nodes within a network \cr
\link[=managedblockchain_list_proposal_votes]{list_proposal_votes} \tab Returns the listing of votes for a specified proposal, including the value of each vote and the unique identifier of the member that cast the vote \cr
\link[=managedblockchain_list_proposals]{list_proposals} \tab Returns a listing of proposals for the network \cr
\link[=managedblockchain_reject_invitation]{reject_invitation} \tab Rejects an invitation to join a network \cr
\link[=managedblockchain_vote_on_proposal]{vote_on_proposal} \tab Casts a vote for a specified ProposalId on behalf of a member
}
}

\examples{
\donttest{svc <- managedblockchain()
svc$create_member(
  Foo = 123
)}

}
