% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_create_member}
\alias{managedblockchain_create_member}
\title{Creates a member within a Managed Blockchain network}
\usage{
managedblockchain_create_member(ClientRequestToken, InvitationId,
  NetworkId, MemberConfiguration)
}
\arguments{
\item{ClientRequestToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the operation. An idempotent operation completes no more
than one time. This identifier is required only if you make a service
request directly using an HTTP client. It is generated automatically if
you use an AWS SDK or the AWS CLI.}

\item{InvitationId}{[required] The unique identifier of the invitation that is sent to the member to
join the network.}

\item{NetworkId}{[required] The unique identifier of the network in which the member is created.}

\item{MemberConfiguration}{[required] Member configuration parameters.}
}
\description{
Creates a member within a Managed Blockchain network.
}
\section{Request syntax}{
\preformatted{svc$create_member(
  ClientRequestToken = "string",
  InvitationId = "string",
  NetworkId = "string",
  MemberConfiguration = list(
    Name = "string",
    Description = "string",
    FrameworkConfiguration = list(
      Fabric = list(
        AdminUsername = "string",
        AdminPassword = "string"
      )
    )
  )
)
}
}

\keyword{internal}
