% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_create_network}
\alias{managedblockchain_create_network}
\title{Creates a new blockchain network using Amazon Managed Blockchain}
\usage{
managedblockchain_create_network(ClientRequestToken, Name, Description,
  Framework, FrameworkVersion, FrameworkConfiguration, VotingPolicy,
  MemberConfiguration)
}
\arguments{
\item{ClientRequestToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the operation. An idempotent operation completes no more
than one time. This identifier is required only if you make a service
request directly using an HTTP client. It is generated automatically if
you use an AWS SDK or the AWS CLI.}

\item{Name}{[required] The name of the network.}

\item{Description}{An optional description for the network.}

\item{Framework}{[required] The blockchain framework that the network uses.}

\item{FrameworkVersion}{[required] The version of the blockchain framework that the network uses.}

\item{FrameworkConfiguration}{Configuration properties of the blockchain framework relevant to the
network configuration.}

\item{VotingPolicy}{[required] The voting rules used by the network to determine if a proposal is
approved.}

\item{MemberConfiguration}{[required] Configuration properties for the first member within the network.}
}
\description{
Creates a new blockchain network using Amazon Managed Blockchain.
}
\section{Request syntax}{
\preformatted{svc$create_network(
  ClientRequestToken = "string",
  Name = "string",
  Description = "string",
  Framework = "HYPERLEDGER_FABRIC",
  FrameworkVersion = "string",
  FrameworkConfiguration = list(
    Fabric = list(
      Edition = "STARTER"|"STANDARD"
    )
  ),
  VotingPolicy = list(
    ApprovalThresholdPolicy = list(
      ThresholdPercentage = 123,
      ProposalDurationInHours = 123,
      ThresholdComparator = "GREATER_THAN"|"GREATER_THAN_OR_EQUAL_TO"
    )
  ),
  MemberConfiguration = list(
    Name = "string",
    Description = "string",
    FrameworkConfiguration = list(
      Fabric = list(
        AdminUsername = "string",
        AdminPassword = "string"
      )
    ),
    LogPublishingConfiguration = list(
      Fabric = list(
        CaLogs = list(
          Cloudwatch = list(
            Enabled = TRUE|FALSE
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
