% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_create_proposal}
\alias{managedblockchain_create_proposal}
\title{Creates a proposal for a change to the network that other members of the
network can vote on, for example, a proposal to add a new member to the
network}
\usage{
managedblockchain_create_proposal(ClientRequestToken, NetworkId,
  MemberId, Actions, Description)
}
\arguments{
\item{ClientRequestToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the operation. An idempotent operation completes no more
than one time. This identifier is required only if you make a service
request directly using an HTTP client. It is generated automatically if
you use an AWS SDK or the AWS CLI.}

\item{NetworkId}{[required] The unique identifier of the network for which the proposal is made.}

\item{MemberId}{[required] The unique identifier of the member that is creating the proposal. This
identifier is especially useful for identifying the member making the
proposal when multiple members exist in a single AWS account.}

\item{Actions}{[required] The type of actions proposed, such as inviting a member or removing a
member. The types of \code{Actions} in a proposal are mutually exclusive. For
example, a proposal with \code{Invitations} actions cannot also contain
\code{Removals} actions.}

\item{Description}{A description for the proposal that is visible to voting members, for
example, "Proposal to add Example Corp. as member."}
}
\description{
Creates a proposal for a change to the network that other members of the
network can vote on, for example, a proposal to add a new member to the
network. Any member can create a proposal.
}
\details{
Applies only to Hyperledger Fabric.
}
\section{Request syntax}{
\preformatted{svc$create_proposal(
  ClientRequestToken = "string",
  NetworkId = "string",
  MemberId = "string",
  Actions = list(
    Invitations = list(
      list(
        Principal = "string"
      )
    ),
    Removals = list(
      list(
        MemberId = "string"
      )
    )
  ),
  Description = "string"
)
}
}

\keyword{internal}
