% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_delete_member}
\alias{managedblockchain_delete_member}
\title{Deletes a member}
\usage{
managedblockchain_delete_member(NetworkId, MemberId)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network from which the member is removed.}

\item{MemberId}{[required] The unique identifier of the member to remove.}
}
\description{
Deletes a member. Deleting a member removes the member and all
associated resources from the network. \code{DeleteMember} can only be called
for a specified \code{MemberId} if the principal performing the action is
associated with the AWS account that owns the member. In all other
cases, the \code{DeleteMember} action is carried out as the result of an
approved proposal to remove a member. If \code{MemberId} is the last member
in a network specified by the last AWS account, the network is deleted
also.

Applies only to Hyperledger Fabric.
}
\section{Request syntax}{
\preformatted{svc$delete_member(
  NetworkId = "string",
  MemberId = "string"
)
}
}

\keyword{internal}
