% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_delete_node}
\alias{managedblockchain_delete_node}
\title{Deletes a peer node from a member that your AWS account owns}
\usage{
managedblockchain_delete_node(NetworkId, MemberId, NodeId)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network that the node belongs to.}

\item{MemberId}{[required] The unique identifier of the member that owns this node.}

\item{NodeId}{[required] The unique identifier of the node.}
}
\description{
Deletes a peer node from a member that your AWS account owns. All data
on the node is lost and cannot be recovered.
}
\section{Request syntax}{
\preformatted{svc$delete_node(
  NetworkId = "string",
  MemberId = "string",
  NodeId = "string"
)
}
}

\keyword{internal}
