% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_get_network}
\alias{managedblockchain_get_network}
\title{Returns detailed information about a network}
\usage{
managedblockchain_get_network(NetworkId)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network to get information about.}
}
\description{
Returns detailed information about a network.
}
\details{
Applies to Hyperledger Fabric and Ethereum.
}
\section{Request syntax}{
\preformatted{svc$get_network(
  NetworkId = "string"
)
}
}

\keyword{internal}
