% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_get_node}
\alias{managedblockchain_get_node}
\title{Returns detailed information about a peer node}
\usage{
managedblockchain_get_node(NetworkId, MemberId, NodeId)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network to which the node belongs.}

\item{MemberId}{[required] The unique identifier of the member that owns the node.}

\item{NodeId}{[required] The unique identifier of the node.}
}
\description{
Returns detailed information about a peer node.
}
\section{Request syntax}{
\preformatted{svc$get_node(
  NetworkId = "string",
  MemberId = "string",
  NodeId = "string"
)
}
}

\keyword{internal}
