% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_get_proposal}
\alias{managedblockchain_get_proposal}
\title{Returns detailed information about a proposal}
\usage{
managedblockchain_get_proposal(NetworkId, ProposalId)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network for which the proposal is made.}

\item{ProposalId}{[required] The unique identifier of the proposal.}
}
\description{
Returns detailed information about a proposal.
}
\section{Request syntax}{
\preformatted{svc$get_proposal(
  NetworkId = "string",
  ProposalId = "string"
)
}
}

\keyword{internal}
