% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_list_members}
\alias{managedblockchain_list_members}
\title{Returns a list of the members in a network and properties of their
configurations}
\usage{
managedblockchain_list_members(NetworkId, Name, Status, IsOwned,
  MaxResults, NextToken)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network for which to list members.}

\item{Name}{The optional name of the member to list.}

\item{Status}{An optional status specifier. If provided, only members currently in
this status are listed.}

\item{IsOwned}{An optional Boolean value. If provided, the request is limited either to
members that the current AWS account owns (\code{true}) or that other AWS
accounts own (\code{false}). If omitted, all members are listed.}

\item{MaxResults}{The maximum number of members to return in the request.}

\item{NextToken}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Returns a list of the members in a network and properties of their
configurations.
}
\details{
Applies only to Hyperledger Fabric.
}
\section{Request syntax}{
\preformatted{svc$list_members(
  NetworkId = "string",
  Name = "string",
  Status = "CREATING"|"AVAILABLE"|"CREATE_FAILED"|"UPDATING"|"DELETING"|"DELETED",
  IsOwned = TRUE|FALSE,
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
