% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_list_networks}
\alias{managedblockchain_list_networks}
\title{Returns information about the networks in which the current AWS account
participates}
\usage{
managedblockchain_list_networks(Name, Framework, Status, MaxResults,
  NextToken)
}
\arguments{
\item{Name}{The name of the network.}

\item{Framework}{An optional framework specifier. If provided, only networks of this
framework type are listed.}

\item{Status}{An optional status specifier. If provided, only networks currently in
this status are listed.

Applies only to Hyperledger Fabric.}

\item{MaxResults}{The maximum number of networks to list.}

\item{NextToken}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Returns information about the networks in which the current AWS account
participates.

Applies to Hyperledger Fabric and Ethereum.
}
\section{Request syntax}{
\preformatted{svc$list_networks(
  Name = "string",
  Framework = "HYPERLEDGER_FABRIC"|"ETHEREUM",
  Status = "CREATING"|"AVAILABLE"|"CREATE_FAILED"|"DELETING"|"DELETED",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
