% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_list_proposal_votes}
\alias{managedblockchain_list_proposal_votes}
\title{Returns the listing of votes for a specified proposal, including the
value of each vote and the unique identifier of the member that cast the
vote}
\usage{
managedblockchain_list_proposal_votes(NetworkId, ProposalId, MaxResults,
  NextToken)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network.}

\item{ProposalId}{[required] The unique identifier of the proposal.}

\item{MaxResults}{The maximum number of votes to return.}

\item{NextToken}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Returns the listing of votes for a specified proposal, including the
value of each vote and the unique identifier of the member that cast the
vote.
}
\section{Request syntax}{
\preformatted{svc$list_proposal_votes(
  NetworkId = "string",
  ProposalId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
