% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_list_proposals}
\alias{managedblockchain_list_proposals}
\title{Returns a listing of proposals for the network}
\usage{
managedblockchain_list_proposals(NetworkId, MaxResults, NextToken)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network.}

\item{MaxResults}{The maximum number of proposals to return.}

\item{NextToken}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Returns a listing of proposals for the network.
}
\section{Request syntax}{
\preformatted{svc$list_proposals(
  NetworkId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
