% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_reject_invitation}
\alias{managedblockchain_reject_invitation}
\title{Rejects an invitation to join a network}
\usage{
managedblockchain_reject_invitation(InvitationId)
}
\arguments{
\item{InvitationId}{[required] The unique identifier of the invitation to reject.}
}
\description{
Rejects an invitation to join a network. This action can be called by a
principal in an AWS account that has received an invitation to create a
member and join a network.
}
\details{
Applies only to Hyperledger Fabric.
}
\section{Request syntax}{
\preformatted{svc$reject_invitation(
  InvitationId = "string"
)
}
}

\keyword{internal}
