% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_describe_change_set}
\alias{marketplacecatalog_describe_change_set}
\title{Provides information about a given change set}
\usage{
marketplacecatalog_describe_change_set(Catalog, ChangeSetId)
}
\arguments{
\item{Catalog}{[required] Required. The catalog related to the request. Fixed value:
\code{AWSMarketplace}}

\item{ChangeSetId}{[required] Required. The unique identifier for the
\code{\link[=marketplacecatalog_start_change_set]{start_change_set}} request that
you want to describe the details for.}
}
\description{
Provides information about a given change set.
}
\section{Request syntax}{
\preformatted{svc$describe_change_set(
  Catalog = "string",
  ChangeSetId = "string"
)
}
}

\keyword{internal}
