% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_describe_entity}
\alias{marketplacecatalog_describe_entity}
\title{Returns the metadata and content of the entity}
\usage{
marketplacecatalog_describe_entity(Catalog, EntityId)
}
\arguments{
\item{Catalog}{[required] Required. The catalog related to the request. Fixed value:
\code{AWSMarketplace}}

\item{EntityId}{[required] Required. The unique ID of the entity to describe.}
}
\description{
Returns the metadata and content of the entity.
}
\section{Request syntax}{
\preformatted{svc$describe_entity(
  Catalog = "string",
  EntityId = "string"
)
}
}

\keyword{internal}
