% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_start_change_set}
\alias{marketplacecatalog_start_change_set}
\title{This operation allows you to request changes in your entities}
\usage{
marketplacecatalog_start_change_set(Catalog, ChangeSet, ChangeSetName,
  ClientRequestToken)
}
\arguments{
\item{Catalog}{[required] The catalog related to the request. Fixed value: \code{AWSMarketplace}}

\item{ChangeSet}{[required] Array of \code{change} object.}

\item{ChangeSetName}{Optional case sensitive string of up to 100 ASCII characters. The change
set name can be used to filter the list of change sets.}

\item{ClientRequestToken}{A unique token to identify the request to ensure idempotency.}
}
\description{
This operation allows you to request changes in your entities.
}
\section{Request syntax}{
\preformatted{svc$start_change_set(
  Catalog = "string",
  ChangeSet = list(
    list(
      ChangeType = "string",
      Entity = list(
        Type = "string",
        Identifier = "string"
      ),
      Details = "string"
    )
  ),
  ChangeSetName = "string",
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
