% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_list_tags_for_resource}
\alias{mediaconnect_list_tags_for_resource}
\title{Lists all tags associated with the resource}
\usage{
mediaconnect_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to view tags for.}
}
\description{
Lists all tags associated with the resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
