% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_get_job}
\alias{mediaconvert_get_job}
\title{Retrieve the JSON for a specific completed transcoding job}
\usage{
mediaconvert_get_job(Id)
}
\arguments{
\item{Id}{[required] the job ID of the job.}
}
\description{
Retrieve the JSON for a specific completed transcoding job.
}
\section{Request syntax}{
\preformatted{mediaconvert$get_job(
  Id = "string"
)
}
}

\keyword{internal}
