% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_list_jobs}
\alias{mediaconvert_list_jobs}
\title{Retrieve a JSON array of up to twenty of your most recently created jobs}
\usage{
mediaconvert_list_jobs(MaxResults, NextToken, Order, Queue, Status)
}
\arguments{
\item{MaxResults}{Optional. Number of jobs, up to twenty, that will be returned at one time.}

\item{NextToken}{Use this string, provided with the response to a previous request, to request the next batch of jobs.}

\item{Order}{}

\item{Queue}{Provide a queue name to get back only jobs from that queue.}

\item{Status}{}
}
\description{
Retrieve a JSON array of up to twenty of your most recently created jobs. This array includes in-process, completed, and errored jobs. This will return the jobs themselves, not just a list of the jobs. To retrieve the twenty next most recent jobs, use the nextToken string returned with the array.
}
\section{Request syntax}{
\preformatted{mediaconvert$list_jobs(
  MaxResults = 123,
  NextToken = "string",
  Order = "ASCENDING"|"DESCENDING",
  Queue = "string",
  Status = "SUBMITTED"|"PROGRESSING"|"COMPLETE"|"CANCELED"|"ERROR"
)
}
}

\keyword{internal}
