% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_list_queues}
\alias{mediaconvert_list_queues}
\title{Retrieve a JSON array of up to twenty of your queues}
\usage{
mediaconvert_list_queues(ListBy, MaxResults, NextToken, Order)
}
\arguments{
\item{ListBy}{Optional. When you request a list of queues, you can choose to list them alphabetically by NAME or chronologically by CREATION_DATE. If you don't specify, the service will list them by creation date.}

\item{MaxResults}{Optional. Number of queues, up to twenty, that will be returned at one time.}

\item{NextToken}{Use this string, provided with the response to a previous request, to request the next batch of queues.}

\item{Order}{When you request lists of resources, you can optionally specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.}
}
\description{
Retrieve a JSON array of up to twenty of your queues. This will return the queues themselves, not just a list of them. To retrieve the next twenty queues, use the nextToken string returned with the array.
}
\section{Request syntax}{
\preformatted{mediaconvert$list_queues(
  ListBy = "NAME"|"CREATION_DATE",
  MaxResults = 123,
  NextToken = "string",
  Order = "ASCENDING"|"DESCENDING"
)
}
}

\keyword{internal}
