% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_update_job_template}
\alias{mediaconvert_update_job_template}
\title{Modify one of your existing job templates}
\usage{
mediaconvert_update_job_template(AccelerationSettings, Category,
  Description, HopDestinations, Name, Priority, Queue, Settings,
  StatusUpdateInterval)
}
\arguments{
\item{AccelerationSettings}{Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.}

\item{Category}{The new category for the job template, if you are changing it.}

\item{Description}{The new description for the job template, if you are changing it.}

\item{HopDestinations}{Optional list of hop destinations.}

\item{Name}{[required] The name of the job template you are modifying}

\item{Priority}{Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.}

\item{Queue}{The new queue for the job template, if you are changing it.}

\item{Settings}{JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.}

\item{StatusUpdateInterval}{Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.}
}
\description{
Modify one of your existing job templates.
}
\section{Request syntax}{
\preformatted{svc$update_job_template(
  AccelerationSettings = list(
    Mode = "DISABLED"|"ENABLED"|"PREFERRED"
  ),
  Category = "string",
  Description = "string",
  HopDestinations = list(
    list(
      Priority = 123,
      Queue = "string",
      WaitMinutes = 123
    )
  ),
  Name = "string",
  Priority = 123,
  Queue = "string",
  Settings = list(
    AdAvailOffset = 123,
    AvailBlanking = list(
      AvailBlankingImage = "string"
    ),
    Esam = list(
      ManifestConfirmConditionNotification = list(
        MccXml = "string"
      ),
      ResponseSignalPreroll = 123,
      SignalProcessingNotification = list(
        SccXml = "string"
      )
    ),
    Inputs = list(
      list(
        AudioSelectorGroups = list(
          list(
            AudioSelectorNames = list(
              "string"
            )
          )
        ),
        AudioSelectors = list(
          list(
            CustomLanguageCode = "string",
            DefaultSelection = "DEFAULT"|"NOT_DEFAULT",
            ExternalAudioFileInput = "string",
            LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
            Offset = 123,
            Pids = list(
              123
            ),
            ProgramSelection = 123,
            RemixSettings = list(
              ChannelMapping = list(
                OutputChannels = list(
                  list(
                    InputChannels = list(
                      123
                    ),
                    InputChannelsFineTune = list(
                      123.0
                    )
                  )
                )
              ),
              ChannelsIn = 123,
              ChannelsOut = 123
            ),
            SelectorType = "PID"|"TRACK"|"LANGUAGE_CODE",
            Tracks = list(
              123
            )
          )
        ),
        CaptionSelectors = list(
          list(
            CustomLanguageCode = "string",
            LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
            SourceSettings = list(
              AncillarySourceSettings = list(
                Convert608To708 = "UPCONVERT"|"DISABLED",
                SourceAncillaryChannelNumber = 123,
                TerminateCaptions = "END_OF_INPUT"|"DISABLED"
              ),
              DvbSubSourceSettings = list(
                Pid = 123
              ),
              EmbeddedSourceSettings = list(
                Convert608To708 = "UPCONVERT"|"DISABLED",
                Source608ChannelNumber = 123,
                Source608TrackNumber = 123,
                TerminateCaptions = "END_OF_INPUT"|"DISABLED"
              ),
              FileSourceSettings = list(
                Convert608To708 = "UPCONVERT"|"DISABLED",
                Framerate = list(
                  FramerateDenominator = 123,
                  FramerateNumerator = 123
                ),
                SourceFile = "string",
                TimeDelta = 123
              ),
              SourceType = "ANCILLARY"|"DVB_SUB"|"EMBEDDED"|"SCTE20"|"SCC"|"TTML"|"STL"|"SRT"|"SMI"|"TELETEXT"|"NULL_SOURCE"|"IMSC",
              TeletextSourceSettings = list(
                PageNumber = "string"
              ),
              TrackSourceSettings = list(
                TrackNumber = 123
              )
            )
          )
        ),
        Crop = list(
          Height = 123,
          Width = 123,
          X = 123,
          Y = 123
        ),
        DeblockFilter = "ENABLED"|"DISABLED",
        DenoiseFilter = "ENABLED"|"DISABLED",
        FilterEnable = "AUTO"|"DISABLE"|"FORCE",
        FilterStrength = 123,
        ImageInserter = list(
          InsertableImages = list(
            list(
              Duration = 123,
              FadeIn = 123,
              FadeOut = 123,
              Height = 123,
              ImageInserterInput = "string",
              ImageX = 123,
              ImageY = 123,
              Layer = 123,
              Opacity = 123,
              StartTime = "string",
              Width = 123
            )
          )
        ),
        InputClippings = list(
          list(
            EndTimecode = "string",
            StartTimecode = "string"
          )
        ),
        InputScanType = "AUTO"|"PSF",
        Position = list(
          Height = 123,
          Width = 123,
          X = 123,
          Y = 123
        ),
        ProgramNumber = 123,
        PsiControl = "IGNORE_PSI"|"USE_PSI",
        TimecodeSource = "EMBEDDED"|"ZEROBASED"|"SPECIFIEDSTART",
        TimecodeStart = "string",
        VideoSelector = list(
          AlphaBehavior = "DISCARD"|"REMAP_TO_LUMA",
          ColorSpace = "FOLLOW"|"REC_601"|"REC_709"|"HDR10"|"HLG_2020",
          ColorSpaceUsage = "FORCE"|"FALLBACK",
          Hdr10Metadata = list(
            BluePrimaryX = 123,
            BluePrimaryY = 123,
            GreenPrimaryX = 123,
            GreenPrimaryY = 123,
            MaxContentLightLevel = 123,
            MaxFrameAverageLightLevel = 123,
            MaxLuminance = 123,
            MinLuminance = 123,
            RedPrimaryX = 123,
            RedPrimaryY = 123,
            WhitePointX = 123,
            WhitePointY = 123
          ),
          Pid = 123,
          ProgramNumber = 123,
          Rotate = "DEGREE_0"|"DEGREES_90"|"DEGREES_180"|"DEGREES_270"|"AUTO"
        )
      )
    ),
    MotionImageInserter = list(
      Framerate = list(
        FramerateDenominator = 123,
        FramerateNumerator = 123
      ),
      Input = "string",
      InsertionMode = "MOV"|"PNG",
      Offset = list(
        ImageX = 123,
        ImageY = 123
      ),
      Playback = "ONCE"|"REPEAT",
      StartTime = "string"
    ),
    NielsenConfiguration = list(
      BreakoutCode = 123,
      DistributorId = "string"
    ),
    NielsenNonLinearWatermark = list(
      ActiveWatermarkProcess = "NAES2_AND_NW"|"CBET"|"NAES2_AND_NW_AND_CBET",
      AdiFilename = "string",
      AssetId = "string",
      AssetName = "string",
      CbetSourceId = "string",
      EpisodeId = "string",
      MetadataDestination = "string",
      SourceId = 123,
      SourceWatermarkStatus = "CLEAN"|"WATERMARKED",
      TicServerUrl = "string",
      UniqueTicPerAudioTrack = "RESERVE_UNIQUE_TICS_PER_TRACK"|"SAME_TICS_PER_TRACK"
    ),
    OutputGroups = list(
      list(
        AutomatedEncodingSettings = list(
          AbrSettings = list(
            MaxAbrBitrate = 123,
            MaxRenditions = 123,
            MinAbrBitrate = 123
          )
        ),
        CustomName = "string",
        Name = "string",
        OutputGroupSettings = list(
          CmafGroupSettings = list(
            AdditionalManifests = list(
              list(
                ManifestNameModifier = "string",
                SelectedOutputs = list(
                  "string"
                )
              )
            ),
            BaseUrl = "string",
            ClientCache = "DISABLED"|"ENABLED",
            CodecSpecification = "RFC_6381"|"RFC_4281",
            Destination = "string",
            DestinationSettings = list(
              S3Settings = list(
                AccessControl = list(
                  CannedAcl = "PUBLIC_READ"|"AUTHENTICATED_READ"|"BUCKET_OWNER_READ"|"BUCKET_OWNER_FULL_CONTROL"
                ),
                Encryption = list(
                  EncryptionType = "SERVER_SIDE_ENCRYPTION_S3"|"SERVER_SIDE_ENCRYPTION_KMS",
                  KmsKeyArn = "string"
                )
              )
            ),
            Encryption = list(
              ConstantInitializationVector = "string",
              EncryptionMethod = "SAMPLE_AES"|"AES_CTR",
              InitializationVectorInManifest = "INCLUDE"|"EXCLUDE",
              SpekeKeyProvider = list(
                CertificateArn = "string",
                DashSignaledSystemIds = list(
                  "string"
                ),
                HlsSignaledSystemIds = list(
                  "string"
                ),
                ResourceId = "string",
                Url = "string"
              ),
              StaticKeyProvider = list(
                KeyFormat = "string",
                KeyFormatVersions = "string",
                StaticKeyValue = "string",
                Url = "string"
              ),
              Type = "SPEKE"|"STATIC_KEY"
            ),
            FragmentLength = 123,
            ManifestCompression = "GZIP"|"NONE",
            ManifestDurationFormat = "FLOATING_POINT"|"INTEGER",
            MinBufferTime = 123,
            MinFinalSegmentLength = 123.0,
            MpdProfile = "MAIN_PROFILE"|"ON_DEMAND_PROFILE",
            SegmentControl = "SINGLE_FILE"|"SEGMENTED_FILES",
            SegmentLength = 123,
            StreamInfResolution = "INCLUDE"|"EXCLUDE",
            WriteDashManifest = "DISABLED"|"ENABLED",
            WriteHlsManifest = "DISABLED"|"ENABLED",
            WriteSegmentTimelineInRepresentation = "ENABLED"|"DISABLED"
          ),
          DashIsoGroupSettings = list(
            AdditionalManifests = list(
              list(
                ManifestNameModifier = "string",
                SelectedOutputs = list(
                  "string"
                )
              )
            ),
            BaseUrl = "string",
            Destination = "string",
            DestinationSettings = list(
              S3Settings = list(
                AccessControl = list(
                  CannedAcl = "PUBLIC_READ"|"AUTHENTICATED_READ"|"BUCKET_OWNER_READ"|"BUCKET_OWNER_FULL_CONTROL"
                ),
                Encryption = list(
                  EncryptionType = "SERVER_SIDE_ENCRYPTION_S3"|"SERVER_SIDE_ENCRYPTION_KMS",
                  KmsKeyArn = "string"
                )
              )
            ),
            Encryption = list(
              PlaybackDeviceCompatibility = "CENC_V1"|"UNENCRYPTED_SEI",
              SpekeKeyProvider = list(
                CertificateArn = "string",
                ResourceId = "string",
                SystemIds = list(
                  "string"
                ),
                Url = "string"
              )
            ),
            FragmentLength = 123,
            HbbtvCompliance = "HBBTV_1_5"|"NONE",
            MinBufferTime = 123,
            MinFinalSegmentLength = 123.0,
            MpdProfile = "MAIN_PROFILE"|"ON_DEMAND_PROFILE",
            SegmentControl = "SINGLE_FILE"|"SEGMENTED_FILES",
            SegmentLength = 123,
            WriteSegmentTimelineInRepresentation = "ENABLED"|"DISABLED"
          ),
          FileGroupSettings = list(
            Destination = "string",
            DestinationSettings = list(
              S3Settings = list(
                AccessControl = list(
                  CannedAcl = "PUBLIC_READ"|"AUTHENTICATED_READ"|"BUCKET_OWNER_READ"|"BUCKET_OWNER_FULL_CONTROL"
                ),
                Encryption = list(
                  EncryptionType = "SERVER_SIDE_ENCRYPTION_S3"|"SERVER_SIDE_ENCRYPTION_KMS",
                  KmsKeyArn = "string"
                )
              )
            )
          ),
          HlsGroupSettings = list(
            AdMarkers = list(
              "ELEMENTAL"|"ELEMENTAL_SCTE35"
            ),
            AdditionalManifests = list(
              list(
                ManifestNameModifier = "string",
                SelectedOutputs = list(
                  "string"
                )
              )
            ),
            AudioOnlyHeader = "INCLUDE"|"EXCLUDE",
            BaseUrl = "string",
            CaptionLanguageMappings = list(
              list(
                CaptionChannel = 123,
                CustomLanguageCode = "string",
                LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
                LanguageDescription = "string"
              )
            ),
            CaptionLanguageSetting = "INSERT"|"OMIT"|"NONE",
            ClientCache = "DISABLED"|"ENABLED",
            CodecSpecification = "RFC_6381"|"RFC_4281",
            Destination = "string",
            DestinationSettings = list(
              S3Settings = list(
                AccessControl = list(
                  CannedAcl = "PUBLIC_READ"|"AUTHENTICATED_READ"|"BUCKET_OWNER_READ"|"BUCKET_OWNER_FULL_CONTROL"
                ),
                Encryption = list(
                  EncryptionType = "SERVER_SIDE_ENCRYPTION_S3"|"SERVER_SIDE_ENCRYPTION_KMS",
                  KmsKeyArn = "string"
                )
              )
            ),
            DirectoryStructure = "SINGLE_DIRECTORY"|"SUBDIRECTORY_PER_STREAM",
            Encryption = list(
              ConstantInitializationVector = "string",
              EncryptionMethod = "AES128"|"SAMPLE_AES",
              InitializationVectorInManifest = "INCLUDE"|"EXCLUDE",
              OfflineEncrypted = "ENABLED"|"DISABLED",
              SpekeKeyProvider = list(
                CertificateArn = "string",
                ResourceId = "string",
                SystemIds = list(
                  "string"
                ),
                Url = "string"
              ),
              StaticKeyProvider = list(
                KeyFormat = "string",
                KeyFormatVersions = "string",
                StaticKeyValue = "string",
                Url = "string"
              ),
              Type = "SPEKE"|"STATIC_KEY"
            ),
            ManifestCompression = "GZIP"|"NONE",
            ManifestDurationFormat = "FLOATING_POINT"|"INTEGER",
            MinFinalSegmentLength = 123.0,
            MinSegmentLength = 123,
            OutputSelection = "MANIFESTS_AND_SEGMENTS"|"SEGMENTS_ONLY",
            ProgramDateTime = "INCLUDE"|"EXCLUDE",
            ProgramDateTimePeriod = 123,
            SegmentControl = "SINGLE_FILE"|"SEGMENTED_FILES",
            SegmentLength = 123,
            SegmentsPerSubdirectory = 123,
            StreamInfResolution = "INCLUDE"|"EXCLUDE",
            TimedMetadataId3Frame = "NONE"|"PRIV"|"TDRL",
            TimedMetadataId3Period = 123,
            TimestampDeltaMilliseconds = 123
          ),
          MsSmoothGroupSettings = list(
            AdditionalManifests = list(
              list(
                ManifestNameModifier = "string",
                SelectedOutputs = list(
                  "string"
                )
              )
            ),
            AudioDeduplication = "COMBINE_DUPLICATE_STREAMS"|"NONE",
            Destination = "string",
            DestinationSettings = list(
              S3Settings = list(
                AccessControl = list(
                  CannedAcl = "PUBLIC_READ"|"AUTHENTICATED_READ"|"BUCKET_OWNER_READ"|"BUCKET_OWNER_FULL_CONTROL"
                ),
                Encryption = list(
                  EncryptionType = "SERVER_SIDE_ENCRYPTION_S3"|"SERVER_SIDE_ENCRYPTION_KMS",
                  KmsKeyArn = "string"
                )
              )
            ),
            Encryption = list(
              SpekeKeyProvider = list(
                CertificateArn = "string",
                ResourceId = "string",
                SystemIds = list(
                  "string"
                ),
                Url = "string"
              )
            ),
            FragmentLength = 123,
            ManifestEncoding = "UTF8"|"UTF16"
          ),
          Type = "HLS_GROUP_SETTINGS"|"DASH_ISO_GROUP_SETTINGS"|"FILE_GROUP_SETTINGS"|"MS_SMOOTH_GROUP_SETTINGS"|"CMAF_GROUP_SETTINGS"
        ),
        Outputs = list(
          list(
            AudioDescriptions = list(
              list(
                AudioChannelTaggingSettings = list(
                  ChannelTag = "L"|"R"|"C"|"LFE"|"LS"|"RS"|"LC"|"RC"|"CS"|"LSD"|"RSD"|"TCS"|"VHL"|"VHC"|"VHR"
                ),
                AudioNormalizationSettings = list(
                  Algorithm = "ITU_BS_1770_1"|"ITU_BS_1770_2"|"ITU_BS_1770_3"|"ITU_BS_1770_4",
                  AlgorithmControl = "CORRECT_AUDIO"|"MEASURE_ONLY",
                  CorrectionGateLevel = 123,
                  LoudnessLogging = "LOG"|"DONT_LOG",
                  PeakCalculation = "TRUE_PEAK"|"NONE",
                  TargetLkfs = 123.0
                ),
                AudioSourceName = "string",
                AudioType = 123,
                AudioTypeControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
                CodecSettings = list(
                  AacSettings = list(
                    AudioDescriptionBroadcasterMix = "BROADCASTER_MIXED_AD"|"NORMAL",
                    Bitrate = 123,
                    CodecProfile = "LC"|"HEV1"|"HEV2",
                    CodingMode = "AD_RECEIVER_MIX"|"CODING_MODE_1_0"|"CODING_MODE_1_1"|"CODING_MODE_2_0"|"CODING_MODE_5_1",
                    RateControlMode = "CBR"|"VBR",
                    RawFormat = "LATM_LOAS"|"NONE",
                    SampleRate = 123,
                    Specification = "MPEG2"|"MPEG4",
                    VbrQuality = "LOW"|"MEDIUM_LOW"|"MEDIUM_HIGH"|"HIGH"
                  ),
                  Ac3Settings = list(
                    Bitrate = 123,
                    BitstreamMode = "COMPLETE_MAIN"|"COMMENTARY"|"DIALOGUE"|"EMERGENCY"|"HEARING_IMPAIRED"|"MUSIC_AND_EFFECTS"|"VISUALLY_IMPAIRED"|"VOICE_OVER",
                    CodingMode = "CODING_MODE_1_0"|"CODING_MODE_1_1"|"CODING_MODE_2_0"|"CODING_MODE_3_2_LFE",
                    Dialnorm = 123,
                    DynamicRangeCompressionProfile = "FILM_STANDARD"|"NONE",
                    LfeFilter = "ENABLED"|"DISABLED",
                    MetadataControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
                    SampleRate = 123
                  ),
                  AiffSettings = list(
                    BitDepth = 123,
                    Channels = 123,
                    SampleRate = 123
                  ),
                  Codec = "AAC"|"MP2"|"MP3"|"WAV"|"AIFF"|"AC3"|"EAC3"|"EAC3_ATMOS"|"VORBIS"|"OPUS"|"PASSTHROUGH",
                  Eac3AtmosSettings = list(
                    Bitrate = 123,
                    BitstreamMode = "COMPLETE_MAIN",
                    CodingMode = "CODING_MODE_9_1_6",
                    DialogueIntelligence = "ENABLED"|"DISABLED",
                    DynamicRangeCompressionLine = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
                    DynamicRangeCompressionRf = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
                    LoRoCenterMixLevel = 123.0,
                    LoRoSurroundMixLevel = 123.0,
                    LtRtCenterMixLevel = 123.0,
                    LtRtSurroundMixLevel = 123.0,
                    MeteringMode = "LEQ_A"|"ITU_BS_1770_1"|"ITU_BS_1770_2"|"ITU_BS_1770_3"|"ITU_BS_1770_4",
                    SampleRate = 123,
                    SpeechThreshold = 123,
                    StereoDownmix = "NOT_INDICATED"|"STEREO"|"SURROUND"|"DPL2",
                    SurroundExMode = "NOT_INDICATED"|"ENABLED"|"DISABLED"
                  ),
                  Eac3Settings = list(
                    AttenuationControl = "ATTENUATE_3_DB"|"NONE",
                    Bitrate = 123,
                    BitstreamMode = "COMPLETE_MAIN"|"COMMENTARY"|"EMERGENCY"|"HEARING_IMPAIRED"|"VISUALLY_IMPAIRED",
                    CodingMode = "CODING_MODE_1_0"|"CODING_MODE_2_0"|"CODING_MODE_3_2",
                    DcFilter = "ENABLED"|"DISABLED",
                    Dialnorm = 123,
                    DynamicRangeCompressionLine = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
                    DynamicRangeCompressionRf = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
                    LfeControl = "LFE"|"NO_LFE",
                    LfeFilter = "ENABLED"|"DISABLED",
                    LoRoCenterMixLevel = 123.0,
                    LoRoSurroundMixLevel = 123.0,
                    LtRtCenterMixLevel = 123.0,
                    LtRtSurroundMixLevel = 123.0,
                    MetadataControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
                    PassthroughControl = "WHEN_POSSIBLE"|"NO_PASSTHROUGH",
                    PhaseControl = "SHIFT_90_DEGREES"|"NO_SHIFT",
                    SampleRate = 123,
                    StereoDownmix = "NOT_INDICATED"|"LO_RO"|"LT_RT"|"DPL2",
                    SurroundExMode = "NOT_INDICATED"|"ENABLED"|"DISABLED",
                    SurroundMode = "NOT_INDICATED"|"ENABLED"|"DISABLED"
                  ),
                  Mp2Settings = list(
                    Bitrate = 123,
                    Channels = 123,
                    SampleRate = 123
                  ),
                  Mp3Settings = list(
                    Bitrate = 123,
                    Channels = 123,
                    RateControlMode = "CBR"|"VBR",
                    SampleRate = 123,
                    VbrQuality = 123
                  ),
                  OpusSettings = list(
                    Bitrate = 123,
                    Channels = 123,
                    SampleRate = 123
                  ),
                  VorbisSettings = list(
                    Channels = 123,
                    SampleRate = 123,
                    VbrQuality = 123
                  ),
                  WavSettings = list(
                    BitDepth = 123,
                    Channels = 123,
                    Format = "RIFF"|"RF64",
                    SampleRate = 123
                  )
                ),
                CustomLanguageCode = "string",
                LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
                LanguageCodeControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
                RemixSettings = list(
                  ChannelMapping = list(
                    OutputChannels = list(
                      list(
                        InputChannels = list(
                          123
                        ),
                        InputChannelsFineTune = list(
                          123.0
                        )
                      )
                    )
                  ),
                  ChannelsIn = 123,
                  ChannelsOut = 123
                ),
                StreamName = "string"
              )
            ),
            CaptionDescriptions = list(
              list(
                CaptionSelectorName = "string",
                CustomLanguageCode = "string",
                DestinationSettings = list(
                  BurninDestinationSettings = list(
                    Alignment = "CENTERED"|"LEFT",
                    BackgroundColor = "NONE"|"BLACK"|"WHITE",
                    BackgroundOpacity = 123,
                    FontColor = "WHITE"|"BLACK"|"YELLOW"|"RED"|"GREEN"|"BLUE",
                    FontOpacity = 123,
                    FontResolution = 123,
                    FontScript = "AUTOMATIC"|"HANS"|"HANT",
                    FontSize = 123,
                    OutlineColor = "BLACK"|"WHITE"|"YELLOW"|"RED"|"GREEN"|"BLUE",
                    OutlineSize = 123,
                    ShadowColor = "NONE"|"BLACK"|"WHITE",
                    ShadowOpacity = 123,
                    ShadowXOffset = 123,
                    ShadowYOffset = 123,
                    TeletextSpacing = "FIXED_GRID"|"PROPORTIONAL",
                    XPosition = 123,
                    YPosition = 123
                  ),
                  DestinationType = "BURN_IN"|"DVB_SUB"|"EMBEDDED"|"EMBEDDED_PLUS_SCTE20"|"IMSC"|"SCTE20_PLUS_EMBEDDED"|"SCC"|"SRT"|"SMI"|"TELETEXT"|"TTML"|"WEBVTT",
                  DvbSubDestinationSettings = list(
                    Alignment = "CENTERED"|"LEFT",
                    BackgroundColor = "NONE"|"BLACK"|"WHITE",
                    BackgroundOpacity = 123,
                    FontColor = "WHITE"|"BLACK"|"YELLOW"|"RED"|"GREEN"|"BLUE",
                    FontOpacity = 123,
                    FontResolution = 123,
                    FontScript = "AUTOMATIC"|"HANS"|"HANT",
                    FontSize = 123,
                    OutlineColor = "BLACK"|"WHITE"|"YELLOW"|"RED"|"GREEN"|"BLUE",
                    OutlineSize = 123,
                    ShadowColor = "NONE"|"BLACK"|"WHITE",
                    ShadowOpacity = 123,
                    ShadowXOffset = 123,
                    ShadowYOffset = 123,
                    SubtitlingType = "HEARING_IMPAIRED"|"STANDARD",
                    TeletextSpacing = "FIXED_GRID"|"PROPORTIONAL",
                    XPosition = 123,
                    YPosition = 123
                  ),
                  EmbeddedDestinationSettings = list(
                    Destination608ChannelNumber = 123,
                    Destination708ServiceNumber = 123
                  ),
                  ImscDestinationSettings = list(
                    StylePassthrough = "ENABLED"|"DISABLED"
                  ),
                  SccDestinationSettings = list(
                    Framerate = "FRAMERATE_23_97"|"FRAMERATE_24"|"FRAMERATE_25"|"FRAMERATE_29_97_DROPFRAME"|"FRAMERATE_29_97_NON_DROPFRAME"
                  ),
                  TeletextDestinationSettings = list(
                    PageNumber = "string",
                    PageTypes = list(
                      "PAGE_TYPE_INITIAL"|"PAGE_TYPE_SUBTITLE"|"PAGE_TYPE_ADDL_INFO"|"PAGE_TYPE_PROGRAM_SCHEDULE"|"PAGE_TYPE_HEARING_IMPAIRED_SUBTITLE"
                    )
                  ),
                  TtmlDestinationSettings = list(
                    StylePassthrough = "ENABLED"|"DISABLED"
                  )
                ),
                LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
                LanguageDescription = "string"
              )
            ),
            ContainerSettings = list(
              CmfcSettings = list(
                AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
                IFrameOnlyManifest = "INCLUDE"|"EXCLUDE",
                Scte35Esam = "INSERT"|"NONE",
                Scte35Source = "PASSTHROUGH"|"NONE"
              ),
              Container = "F4V"|"ISMV"|"M2TS"|"M3U8"|"CMFC"|"MOV"|"MP4"|"MPD"|"MXF"|"WEBM"|"RAW",
              F4vSettings = list(
                MoovPlacement = "PROGRESSIVE_DOWNLOAD"|"NORMAL"
              ),
              M2tsSettings = list(
                AudioBufferModel = "DVB"|"ATSC",
                AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
                AudioFramesPerPes = 123,
                AudioPids = list(
                  123
                ),
                Bitrate = 123,
                BufferModel = "MULTIPLEX"|"NONE",
                DvbNitSettings = list(
                  NetworkId = 123,
                  NetworkName = "string",
                  NitInterval = 123
                ),
                DvbSdtSettings = list(
                  OutputSdt = "SDT_FOLLOW"|"SDT_FOLLOW_IF_PRESENT"|"SDT_MANUAL"|"SDT_NONE",
                  SdtInterval = 123,
                  ServiceName = "string",
                  ServiceProviderName = "string"
                ),
                DvbSubPids = list(
                  123
                ),
                DvbTdtSettings = list(
                  TdtInterval = 123
                ),
                DvbTeletextPid = 123,
                EbpAudioInterval = "VIDEO_AND_FIXED_INTERVALS"|"VIDEO_INTERVAL",
                EbpPlacement = "VIDEO_AND_AUDIO_PIDS"|"VIDEO_PID",
                EsRateInPes = "INCLUDE"|"EXCLUDE",
                ForceTsVideoEbpOrder = "FORCE"|"DEFAULT",
                FragmentTime = 123.0,
                MaxPcrInterval = 123,
                MinEbpInterval = 123,
                NielsenId3 = "INSERT"|"NONE",
                NullPacketBitrate = 123.0,
                PatInterval = 123,
                PcrControl = "PCR_EVERY_PES_PACKET"|"CONFIGURED_PCR_PERIOD",
                PcrPid = 123,
                PmtInterval = 123,
                PmtPid = 123,
                PrivateMetadataPid = 123,
                ProgramNumber = 123,
                RateMode = "VBR"|"CBR",
                Scte35Esam = list(
                  Scte35EsamPid = 123
                ),
                Scte35Pid = 123,
                Scte35Source = "PASSTHROUGH"|"NONE",
                SegmentationMarkers = "NONE"|"RAI_SEGSTART"|"RAI_ADAPT"|"PSI_SEGSTART"|"EBP"|"EBP_LEGACY",
                SegmentationStyle = "MAINTAIN_CADENCE"|"RESET_CADENCE",
                SegmentationTime = 123.0,
                TimedMetadataPid = 123,
                TransportStreamId = 123,
                VideoPid = 123
              ),
              M3u8Settings = list(
                AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
                AudioFramesPerPes = 123,
                AudioPids = list(
                  123
                ),
                NielsenId3 = "INSERT"|"NONE",
                PatInterval = 123,
                PcrControl = "PCR_EVERY_PES_PACKET"|"CONFIGURED_PCR_PERIOD",
                PcrPid = 123,
                PmtInterval = 123,
                PmtPid = 123,
                PrivateMetadataPid = 123,
                ProgramNumber = 123,
                Scte35Pid = 123,
                Scte35Source = "PASSTHROUGH"|"NONE",
                TimedMetadata = "PASSTHROUGH"|"NONE",
                TimedMetadataPid = 123,
                TransportStreamId = 123,
                VideoPid = 123
              ),
              MovSettings = list(
                ClapAtom = "INCLUDE"|"EXCLUDE",
                CslgAtom = "INCLUDE"|"EXCLUDE",
                Mpeg2FourCCControl = "XDCAM"|"MPEG",
                PaddingControl = "OMNEON"|"NONE",
                Reference = "SELF_CONTAINED"|"EXTERNAL"
              ),
              Mp4Settings = list(
                AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
                CslgAtom = "INCLUDE"|"EXCLUDE",
                CttsVersion = 123,
                FreeSpaceBox = "INCLUDE"|"EXCLUDE",
                MoovPlacement = "PROGRESSIVE_DOWNLOAD"|"NORMAL",
                Mp4MajorBrand = "string"
              ),
              MpdSettings = list(
                AccessibilityCaptionHints = "INCLUDE"|"EXCLUDE",
                AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
                CaptionContainerType = "RAW"|"FRAGMENTED_MP4",
                Scte35Esam = "INSERT"|"NONE",
                Scte35Source = "PASSTHROUGH"|"NONE"
              ),
              MxfSettings = list(
                AfdSignaling = "NO_COPY"|"COPY_FROM_VIDEO",
                Profile = "D_10"|"XDCAM"|"OP1A"
              )
            ),
            Extension = "string",
            NameModifier = "string",
            OutputSettings = list(
              HlsSettings = list(
                AudioGroupId = "string",
                AudioOnlyContainer = "AUTOMATIC"|"M2TS",
                AudioRenditionSets = "string",
                AudioTrackType = "ALTERNATE_AUDIO_AUTO_SELECT_DEFAULT"|"ALTERNATE_AUDIO_AUTO_SELECT"|"ALTERNATE_AUDIO_NOT_AUTO_SELECT"|"AUDIO_ONLY_VARIANT_STREAM",
                IFrameOnlyManifest = "INCLUDE"|"EXCLUDE",
                SegmentModifier = "string"
              )
            ),
            Preset = "string",
            VideoDescription = list(
              AfdSignaling = "NONE"|"AUTO"|"FIXED",
              AntiAlias = "DISABLED"|"ENABLED",
              CodecSettings = list(
                Av1Settings = list(
                  AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
                  FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  GopSize = 123.0,
                  MaxBitrate = 123,
                  NumberBFramesBetweenReferenceFrames = 123,
                  QvbrSettings = list(
                    QvbrQualityLevel = 123,
                    QvbrQualityLevelFineTune = 123.0
                  ),
                  RateControlMode = "QVBR",
                  Slices = 123,
                  SpatialAdaptiveQuantization = "DISABLED"|"ENABLED"
                ),
                AvcIntraSettings = list(
                  AvcIntraClass = "CLASS_50"|"CLASS_100"|"CLASS_200"|"CLASS_4K_2K",
                  AvcIntraUhdSettings = list(
                    QualityTuningLevel = "SINGLE_PASS"|"MULTI_PASS"
                  ),
                  FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
                  ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
                  SlowPal = "DISABLED"|"ENABLED",
                  Telecine = "NONE"|"HARD"
                ),
                Codec = "AV1"|"AVC_INTRA"|"FRAME_CAPTURE"|"H_264"|"H_265"|"MPEG2"|"PRORES"|"VC3"|"VP8"|"VP9",
                FrameCaptureSettings = list(
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  MaxCaptures = 123,
                  Quality = 123
                ),
                H264Settings = list(
                  AdaptiveQuantization = "OFF"|"AUTO"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
                  Bitrate = 123,
                  CodecLevel = "AUTO"|"LEVEL_1"|"LEVEL_1_1"|"LEVEL_1_2"|"LEVEL_1_3"|"LEVEL_2"|"LEVEL_2_1"|"LEVEL_2_2"|"LEVEL_3"|"LEVEL_3_1"|"LEVEL_3_2"|"LEVEL_4"|"LEVEL_4_1"|"LEVEL_4_2"|"LEVEL_5"|"LEVEL_5_1"|"LEVEL_5_2",
                  CodecProfile = "BASELINE"|"HIGH"|"HIGH_10BIT"|"HIGH_422"|"HIGH_422_10BIT"|"MAIN",
                  DynamicSubGop = "ADAPTIVE"|"STATIC",
                  EntropyEncoding = "CABAC"|"CAVLC",
                  FieldEncoding = "PAFF"|"FORCE_FIELD",
                  FlickerAdaptiveQuantization = "DISABLED"|"ENABLED",
                  FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  GopBReference = "DISABLED"|"ENABLED",
                  GopClosedCadence = 123,
                  GopSize = 123.0,
                  GopSizeUnits = "FRAMES"|"SECONDS",
                  HrdBufferInitialFillPercentage = 123,
                  HrdBufferSize = 123,
                  InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
                  MaxBitrate = 123,
                  MinIInterval = 123,
                  NumberBFramesBetweenReferenceFrames = 123,
                  NumberReferenceFrames = 123,
                  ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  ParDenominator = 123,
                  ParNumerator = 123,
                  QualityTuningLevel = "SINGLE_PASS"|"SINGLE_PASS_HQ"|"MULTI_PASS_HQ",
                  QvbrSettings = list(
                    MaxAverageBitrate = 123,
                    QvbrQualityLevel = 123,
                    QvbrQualityLevelFineTune = 123.0
                  ),
                  RateControlMode = "VBR"|"CBR"|"QVBR",
                  RepeatPps = "DISABLED"|"ENABLED",
                  ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
                  SceneChangeDetect = "DISABLED"|"ENABLED"|"TRANSITION_DETECTION",
                  Slices = 123,
                  SlowPal = "DISABLED"|"ENABLED",
                  Softness = 123,
                  SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
                  Syntax = "DEFAULT"|"RP2027",
                  Telecine = "NONE"|"SOFT"|"HARD",
                  TemporalAdaptiveQuantization = "DISABLED"|"ENABLED",
                  UnregisteredSeiTimecode = "DISABLED"|"ENABLED"
                ),
                H265Settings = list(
                  AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
                  AlternateTransferFunctionSei = "DISABLED"|"ENABLED",
                  Bitrate = 123,
                  CodecLevel = "AUTO"|"LEVEL_1"|"LEVEL_2"|"LEVEL_2_1"|"LEVEL_3"|"LEVEL_3_1"|"LEVEL_4"|"LEVEL_4_1"|"LEVEL_5"|"LEVEL_5_1"|"LEVEL_5_2"|"LEVEL_6"|"LEVEL_6_1"|"LEVEL_6_2",
                  CodecProfile = "MAIN_MAIN"|"MAIN_HIGH"|"MAIN10_MAIN"|"MAIN10_HIGH"|"MAIN_422_8BIT_MAIN"|"MAIN_422_8BIT_HIGH"|"MAIN_422_10BIT_MAIN"|"MAIN_422_10BIT_HIGH",
                  DynamicSubGop = "ADAPTIVE"|"STATIC",
                  FlickerAdaptiveQuantization = "DISABLED"|"ENABLED",
                  FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  GopBReference = "DISABLED"|"ENABLED",
                  GopClosedCadence = 123,
                  GopSize = 123.0,
                  GopSizeUnits = "FRAMES"|"SECONDS",
                  HrdBufferInitialFillPercentage = 123,
                  HrdBufferSize = 123,
                  InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
                  MaxBitrate = 123,
                  MinIInterval = 123,
                  NumberBFramesBetweenReferenceFrames = 123,
                  NumberReferenceFrames = 123,
                  ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  ParDenominator = 123,
                  ParNumerator = 123,
                  QualityTuningLevel = "SINGLE_PASS"|"SINGLE_PASS_HQ"|"MULTI_PASS_HQ",
                  QvbrSettings = list(
                    MaxAverageBitrate = 123,
                    QvbrQualityLevel = 123,
                    QvbrQualityLevelFineTune = 123.0
                  ),
                  RateControlMode = "VBR"|"CBR"|"QVBR",
                  SampleAdaptiveOffsetFilterMode = "DEFAULT"|"ADAPTIVE"|"OFF",
                  ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
                  SceneChangeDetect = "DISABLED"|"ENABLED"|"TRANSITION_DETECTION",
                  Slices = 123,
                  SlowPal = "DISABLED"|"ENABLED",
                  SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
                  Telecine = "NONE"|"SOFT"|"HARD",
                  TemporalAdaptiveQuantization = "DISABLED"|"ENABLED",
                  TemporalIds = "DISABLED"|"ENABLED",
                  Tiles = "DISABLED"|"ENABLED",
                  UnregisteredSeiTimecode = "DISABLED"|"ENABLED",
                  WriteMp4PackagingType = "HVC1"|"HEV1"
                ),
                Mpeg2Settings = list(
                  AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH",
                  Bitrate = 123,
                  CodecLevel = "AUTO"|"LOW"|"MAIN"|"HIGH1440"|"HIGH",
                  CodecProfile = "MAIN"|"PROFILE_422",
                  DynamicSubGop = "ADAPTIVE"|"STATIC",
                  FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  GopClosedCadence = 123,
                  GopSize = 123.0,
                  GopSizeUnits = "FRAMES"|"SECONDS",
                  HrdBufferInitialFillPercentage = 123,
                  HrdBufferSize = 123,
                  InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
                  IntraDcPrecision = "AUTO"|"INTRA_DC_PRECISION_8"|"INTRA_DC_PRECISION_9"|"INTRA_DC_PRECISION_10"|"INTRA_DC_PRECISION_11",
                  MaxBitrate = 123,
                  MinIInterval = 123,
                  NumberBFramesBetweenReferenceFrames = 123,
                  ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  ParDenominator = 123,
                  ParNumerator = 123,
                  QualityTuningLevel = "SINGLE_PASS"|"MULTI_PASS",
                  RateControlMode = "VBR"|"CBR",
                  ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
                  SceneChangeDetect = "DISABLED"|"ENABLED",
                  SlowPal = "DISABLED"|"ENABLED",
                  Softness = 123,
                  SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
                  Syntax = "DEFAULT"|"D_10",
                  Telecine = "NONE"|"SOFT"|"HARD",
                  TemporalAdaptiveQuantization = "DISABLED"|"ENABLED"
                ),
                ProresSettings = list(
                  CodecProfile = "APPLE_PRORES_422"|"APPLE_PRORES_422_HQ"|"APPLE_PRORES_422_LT"|"APPLE_PRORES_422_PROXY",
                  FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
                  ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  ParDenominator = 123,
                  ParNumerator = 123,
                  ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
                  SlowPal = "DISABLED"|"ENABLED",
                  Telecine = "NONE"|"HARD"
                ),
                Vc3Settings = list(
                  FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  InterlaceMode = "INTERLACED"|"PROGRESSIVE",
                  ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
                  SlowPal = "DISABLED"|"ENABLED",
                  Telecine = "NONE"|"HARD",
                  Vc3Class = "CLASS_145_8BIT"|"CLASS_220_8BIT"|"CLASS_220_10BIT"
                ),
                Vp8Settings = list(
                  Bitrate = 123,
                  FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  GopSize = 123.0,
                  HrdBufferSize = 123,
                  MaxBitrate = 123,
                  ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  ParDenominator = 123,
                  ParNumerator = 123,
                  QualityTuningLevel = "MULTI_PASS"|"MULTI_PASS_HQ",
                  RateControlMode = "VBR"
                ),
                Vp9Settings = list(
                  Bitrate = 123,
                  FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
                  FramerateDenominator = 123,
                  FramerateNumerator = 123,
                  GopSize = 123.0,
                  HrdBufferSize = 123,
                  MaxBitrate = 123,
                  ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
                  ParDenominator = 123,
                  ParNumerator = 123,
                  QualityTuningLevel = "MULTI_PASS"|"MULTI_PASS_HQ",
                  RateControlMode = "VBR"
                )
              ),
              ColorMetadata = "IGNORE"|"INSERT",
              Crop = list(
                Height = 123,
                Width = 123,
                X = 123,
                Y = 123
              ),
              DropFrameTimecode = "DISABLED"|"ENABLED",
              FixedAfd = 123,
              Height = 123,
              Position = list(
                Height = 123,
                Width = 123,
                X = 123,
                Y = 123
              ),
              RespondToAfd = "NONE"|"RESPOND"|"PASSTHROUGH",
              ScalingBehavior = "DEFAULT"|"STRETCH_TO_OUTPUT",
              Sharpness = 123,
              TimecodeInsertion = "DISABLED"|"PIC_TIMING_SEI",
              VideoPreprocessors = list(
                ColorCorrector = list(
                  Brightness = 123,
                  ColorSpaceConversion = "NONE"|"FORCE_601"|"FORCE_709"|"FORCE_HDR10"|"FORCE_HLG_2020",
                  Contrast = 123,
                  Hdr10Metadata = list(
                    BluePrimaryX = 123,
                    BluePrimaryY = 123,
                    GreenPrimaryX = 123,
                    GreenPrimaryY = 123,
                    MaxContentLightLevel = 123,
                    MaxFrameAverageLightLevel = 123,
                    MaxLuminance = 123,
                    MinLuminance = 123,
                    RedPrimaryX = 123,
                    RedPrimaryY = 123,
                    WhitePointX = 123,
                    WhitePointY = 123
                  ),
                  Hue = 123,
                  Saturation = 123
                ),
                Deinterlacer = list(
                  Algorithm = "INTERPOLATE"|"INTERPOLATE_TICKER"|"BLEND"|"BLEND_TICKER",
                  Control = "FORCE_ALL_FRAMES"|"NORMAL",
                  Mode = "DEINTERLACE"|"INVERSE_TELECINE"|"ADAPTIVE"
                ),
                DolbyVision = list(
                  L6Metadata = list(
                    MaxCll = 123,
                    MaxFall = 123
                  ),
                  L6Mode = "PASSTHROUGH"|"RECALCULATE"|"SPECIFY",
                  Profile = "PROFILE_5"
                ),
                ImageInserter = list(
                  InsertableImages = list(
                    list(
                      Duration = 123,
                      FadeIn = 123,
                      FadeOut = 123,
                      Height = 123,
                      ImageInserterInput = "string",
                      ImageX = 123,
                      ImageY = 123,
                      Layer = 123,
                      Opacity = 123,
                      StartTime = "string",
                      Width = 123
                    )
                  )
                ),
                NoiseReducer = list(
                  Filter = "BILATERAL"|"MEAN"|"GAUSSIAN"|"LANCZOS"|"SHARPEN"|"CONSERVE"|"SPATIAL"|"TEMPORAL",
                  FilterSettings = list(
                    Strength = 123
                  ),
                  SpatialFilterSettings = list(
                    PostFilterSharpenStrength = 123,
                    Speed = 123,
                    Strength = 123
                  ),
                  TemporalFilterSettings = list(
                    AggressiveMode = 123,
                    PostTemporalSharpening = "DISABLED"|"ENABLED"|"AUTO",
                    Speed = 123,
                    Strength = 123
                  )
                ),
                PartnerWatermarking = list(
                  NexguardFileMarkerSettings = list(
                    License = "string",
                    Payload = 123,
                    Preset = "string",
                    Strength = "LIGHTEST"|"LIGHTER"|"DEFAULT"|"STRONGER"|"STRONGEST"
                  )
                ),
                TimecodeBurnin = list(
                  FontSize = 123,
                  Position = "TOP_CENTER"|"TOP_LEFT"|"TOP_RIGHT"|"MIDDLE_LEFT"|"MIDDLE_CENTER"|"MIDDLE_RIGHT"|"BOTTOM_LEFT"|"BOTTOM_CENTER"|"BOTTOM_RIGHT",
                  Prefix = "string"
                )
              ),
              Width = 123
            )
          )
        )
      )
    ),
    TimecodeConfig = list(
      Anchor = "string",
      Source = "EMBEDDED"|"ZEROBASED"|"SPECIFIEDSTART",
      Start = "string",
      TimestampOffset = "string"
    ),
    TimedMetadataInsertion = list(
      Id3Insertions = list(
        list(
          Id3 = "string",
          Timecode = "string"
        )
      )
    )
  ),
  StatusUpdateInterval = "SECONDS_10"|"SECONDS_12"|"SECONDS_15"|"SECONDS_20"|"SECONDS_30"|"SECONDS_60"|"SECONDS_120"|"SECONDS_180"|"SECONDS_240"|"SECONDS_300"|"SECONDS_360"|"SECONDS_420"|"SECONDS_480"|"SECONDS_540"|"SECONDS_600"
)
}
}

\keyword{internal}
