% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_batch_stop}
\alias{medialive_batch_stop}
\title{Stops running resources}
\usage{
medialive_batch_stop(ChannelIds, MultiplexIds)
}
\arguments{
\item{ChannelIds}{List of channel IDs}

\item{MultiplexIds}{List of multiplex IDs}
}
\description{
Stops running resources
}
\section{Request syntax}{
\preformatted{svc$batch_stop(
  ChannelIds = list(
    "string"
  ),
  MultiplexIds = list(
    "string"
  )
)
}
}

\keyword{internal}
