% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_create_input}
\alias{medialive_create_input}
\title{Create an input}
\usage{
medialive_create_input(Destinations, InputSecurityGroups,
  MediaConnectFlows, Name, RequestId, RoleArn, Sources, Tags, Type, Vpc)
}
\arguments{
\item{Destinations}{Destination settings for PUSH type inputs.}

\item{InputSecurityGroups}{A list of security groups referenced by IDs to attach to the input.}

\item{MediaConnectFlows}{A list of the MediaConnect Flows that you want to use in this input. You can specify as few as one
Flow and presently, as many as two. The only requirement is when you have more than one is that each Flow is in a
separate Availability Zone as this ensures your EML input is redundant to AZ issues.}

\item{Name}{Name of the input.}

\item{RequestId}{Unique identifier of the request to ensure the request is handled
exactly once in case of retries.}

\item{RoleArn}{The Amazon Resource Name (ARN) of the role this input assumes during and after creation.}

\item{Sources}{The source URLs for a PULL-type input. Every PULL type input needs
exactly two source URLs for redundancy.
Only specify sources for PULL type Inputs. Leave Destinations empty.}

\item{Tags}{A collection of key-value pairs.}

\item{Type}{}

\item{Vpc}{}
}
\description{
Create an input
}
\section{Request syntax}{
\preformatted{medialive$create_input(
  Destinations = list(
    list(
      StreamName = "string"
    )
  ),
  InputSecurityGroups = list(
    "string"
  ),
  MediaConnectFlows = list(
    list(
      FlowArn = "string"
    )
  ),
  Name = "string",
  RequestId = "string",
  RoleArn = "string",
  Sources = list(
    list(
      PasswordParam = "string",
      Url = "string",
      Username = "string"
    )
  ),
  Tags = list(
    "string"
  ),
  Type = "UDP_PUSH"|"RTP_PUSH"|"RTMP_PUSH"|"RTMP_PULL"|"URL_PULL"|"MP4_FILE"|"MEDIACONNECT",
  Vpc = list(
    SecurityGroupIds = list(
      "string"
    ),
    SubnetIds = list(
      "string"
    )
  )
)
}
}

\keyword{internal}
