% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_create_multiplex_program}
\alias{medialive_create_multiplex_program}
\title{Create a new program in the multiplex}
\usage{
medialive_create_multiplex_program(MultiplexId,
  MultiplexProgramSettings, ProgramName, RequestId)
}
\arguments{
\item{MultiplexId}{[required] ID of the multiplex where the program is to be created.}

\item{MultiplexProgramSettings}{[required] The settings for this multiplex program.}

\item{ProgramName}{[required] Name of multiplex program.}

\item{RequestId}{[required] Unique request ID. This prevents retries from creating multiple
resources.}
}
\description{
Create a new program in the multiplex.
}
\section{Request syntax}{
\preformatted{svc$create_multiplex_program(
  MultiplexId = "string",
  MultiplexProgramSettings = list(
    ProgramNumber = 123,
    ServiceDescriptor = list(
      ProviderName = "string",
      ServiceName = "string"
    ),
    VideoSettings = list(
      ConstantBitrate = 123,
      StatmuxSettings = list(
        MaximumBitrate = 123,
        MinimumBitrate = 123
      )
    )
  ),
  ProgramName = "string",
  RequestId = "string"
)
}
}

\keyword{internal}
