% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_create_tags}
\alias{medialive_create_tags}
\title{Create tags for a resource}
\usage{
medialive_create_tags(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required]}

\item{Tags}{}
}
\description{
Create tags for a resource
}
\section{Request syntax}{
\preformatted{svc$create_tags(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
