% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_input_device}
\alias{medialive_describe_input_device}
\title{Gets the details for the input device}
\usage{
medialive_describe_input_device(InputDeviceId)
}
\arguments{
\item{InputDeviceId}{[required] The unique ID of this input device. For example, hd-123456789abcdef.}
}
\description{
Gets the details for the input device
}
\section{Request syntax}{
\preformatted{svc$describe_input_device(
  InputDeviceId = "string"
)
}
}

\keyword{internal}
