% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_reservation}
\alias{medialive_describe_reservation}
\title{Get details for a reservation}
\usage{
medialive_describe_reservation(ReservationId)
}
\arguments{
\item{ReservationId}{[required] Unique reservation ID, e.g. '1234567'}
}
\description{
Get details for a reservation.
}
\section{Request syntax}{
\preformatted{svc$describe_reservation(
  ReservationId = "string"
)
}
}

\keyword{internal}
