% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_input_devices}
\alias{medialive_list_input_devices}
\title{List input devices}
\usage{
medialive_list_input_devices(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
List input devices
}
\section{Request syntax}{
\preformatted{svc$list_input_devices(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
