% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_inputs}
\alias{medialive_list_inputs}
\title{Produces list of inputs that have been created}
\usage{
medialive_list_inputs(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Produces list of inputs that have been created
}
\section{Request syntax}{
\preformatted{svc$list_inputs(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
