% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_multiplex_programs}
\alias{medialive_list_multiplex_programs}
\title{List the programs that currently exist for a specific multiplex}
\usage{
medialive_list_multiplex_programs(MaxResults, MultiplexId, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of items to return.}

\item{MultiplexId}{[required] The ID of the multiplex that the programs belong to.}

\item{NextToken}{The token to retrieve the next page of results.}
}
\description{
List the programs that currently exist for a specific multiplex.
}
\section{Request syntax}{
\preformatted{svc$list_multiplex_programs(
  MaxResults = 123,
  MultiplexId = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
