% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_reservations}
\alias{medialive_list_reservations}
\title{List purchased reservations}
\usage{
medialive_list_reservations(ChannelClass, Codec, MaxResults,
  MaximumBitrate, MaximumFramerate, NextToken, Resolution, ResourceType,
  SpecialFeature, VideoQuality)
}
\arguments{
\item{ChannelClass}{Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'}

\item{Codec}{Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'}

\item{MaxResults}{}

\item{MaximumBitrate}{Filter by bitrate, 'MAX_10_MBPS', 'MAX_20_MBPS', or 'MAX_50_MBPS'}

\item{MaximumFramerate}{Filter by framerate, 'MAX_30_FPS' or 'MAX_60_FPS'}

\item{NextToken}{}

\item{Resolution}{Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'}

\item{ResourceType}{Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'}

\item{SpecialFeature}{Filter by special feature, 'ADVANCED_AUDIO' or 'AUDIO_NORMALIZATION'}

\item{VideoQuality}{Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'}
}
\description{
List purchased reservations.
}
\section{Request syntax}{
\preformatted{svc$list_reservations(
  ChannelClass = "string",
  Codec = "string",
  MaxResults = 123,
  MaximumBitrate = "string",
  MaximumFramerate = "string",
  NextToken = "string",
  Resolution = "string",
  ResourceType = "string",
  SpecialFeature = "string",
  VideoQuality = "string"
)
}
}

\keyword{internal}
