% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_tags_for_resource}
\alias{medialive_list_tags_for_resource}
\title{Produces list of tags that have been created for a resource}
\usage{
medialive_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required]}
}
\description{
Produces list of tags that have been created for a resource
}
\section{Request syntax}{
\preformatted{medialive$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
