% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_purchase_offering}
\alias{medialive_purchase_offering}
\title{Purchase an offering and create a reservation}
\usage{
medialive_purchase_offering(Count, Name, OfferingId, RequestId, Start)
}
\arguments{
\item{Count}{[required] Number of resources}

\item{Name}{Name for the new reservation}

\item{OfferingId}{[required] Offering to purchase, e.g. '87654321'}

\item{RequestId}{Unique request ID to be specified. This is needed to prevent retries from creating multiple resources.}

\item{Start}{Requested reservation start time (UTC) in ISO-8601 format. The specified time must be between the first day of the current month and one year from now. If no value is given, the default is now.}
}
\description{
Purchase an offering and create a reservation.
}
\section{Request syntax}{
\preformatted{svc$purchase_offering(
  Count = 123,
  Name = "string",
  OfferingId = "string",
  RequestId = "string",
  Start = "string"
)
}
}

\keyword{internal}
