% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_stop_channel}
\alias{medialive_stop_channel}
\title{Stops a running channel}
\usage{
medialive_stop_channel(ChannelId)
}
\arguments{
\item{ChannelId}{[required] A request to stop a running channel}
}
\description{
Stops a running channel
}
\section{Request syntax}{
\preformatted{medialive$stop_channel(
  ChannelId = "string"
)
}
}

\keyword{internal}
