% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_stop_multiplex}
\alias{medialive_stop_multiplex}
\title{Stops a running multiplex}
\usage{
medialive_stop_multiplex(MultiplexId)
}
\arguments{
\item{MultiplexId}{[required] The ID of the multiplex.}
}
\description{
Stops a running multiplex. If the multiplex isn't running, this action has no effect.
}
\section{Request syntax}{
\preformatted{svc$stop_multiplex(
  MultiplexId = "string"
)
}
}

\keyword{internal}
