% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_update_channel_class}
\alias{medialive_update_channel_class}
\title{Changes the class of the channel}
\usage{
medialive_update_channel_class(ChannelClass, ChannelId, Destinations)
}
\arguments{
\item{ChannelClass}{[required] The channel class that you wish to update this channel to use.}

\item{ChannelId}{[required] Channel Id of the channel whose class should be updated.}

\item{Destinations}{A list of output destinations for this channel.}
}
\description{
Changes the class of the channel.
}
\section{Request syntax}{
\preformatted{svc$update_channel_class(
  ChannelClass = "STANDARD"|"SINGLE_PIPELINE",
  ChannelId = "string",
  Destinations = list(
    list(
      Id = "string",
      MediaPackageSettings = list(
        list(
          ChannelId = "string"
        )
      ),
      MultiplexSettings = list(
        MultiplexId = "string",
        ProgramName = "string"
      ),
      Settings = list(
        list(
          PasswordParam = "string",
          StreamName = "string",
          Url = "string",
          Username = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
