% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_update_input_security_group}
\alias{medialive_update_input_security_group}
\title{Update an Input Security Group's Whilelists}
\usage{
medialive_update_input_security_group(InputSecurityGroupId, Tags,
  WhitelistRules)
}
\arguments{
\item{InputSecurityGroupId}{[required] The id of the Input Security Group to update.}

\item{Tags}{A collection of key-value pairs.}

\item{WhitelistRules}{List of IPv4 CIDR addresses to whitelist}
}
\description{
Update an Input Security Group's Whilelists.
}
\section{Request syntax}{
\preformatted{medialive$update_input_security_group(
  InputSecurityGroupId = "string",
  Tags = list(
    "string"
  ),
  WhitelistRules = list(
    list(
      Cidr = "string"
    )
  )
)
}
}

\keyword{internal}
