% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_update_multiplex}
\alias{medialive_update_multiplex}
\title{Updates a multiplex}
\usage{
medialive_update_multiplex(MultiplexId, MultiplexSettings, Name)
}
\arguments{
\item{MultiplexId}{[required] ID of the multiplex to update.}

\item{MultiplexSettings}{The new settings for a multiplex.}

\item{Name}{Name of the multiplex.}
}
\description{
Updates a multiplex.
}
\section{Request syntax}{
\preformatted{svc$update_multiplex(
  MultiplexId = "string",
  MultiplexSettings = list(
    MaximumVideoBufferDelayMilliseconds = 123,
    TransportStreamBitrate = 123,
    TransportStreamId = 123,
    TransportStreamReservedBitrate = 123
  ),
  Name = "string"
)
}
}

\keyword{internal}
