% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_update_reservation}
\alias{medialive_update_reservation}
\title{Update reservation}
\usage{
medialive_update_reservation(Name, ReservationId)
}
\arguments{
\item{Name}{Name of the reservation}

\item{ReservationId}{[required] Unique reservation ID, e.g. '1234567'}
}
\description{
Update reservation.
}
\section{Request syntax}{
\preformatted{svc$update_reservation(
  Name = "string",
  ReservationId = "string"
)
}
}

\keyword{internal}
