% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_create_harvest_job}
\alias{mediapackage_create_harvest_job}
\title{Creates a new HarvestJob record}
\usage{
mediapackage_create_harvest_job(EndTime, Id, OriginEndpointId,
  S3Destination, StartTime)
}
\arguments{
\item{EndTime}{[required] The end of the time-window which will be harvested}

\item{Id}{[required] The ID of the HarvestJob. The ID must be unique within the region
and it cannot be changed after the HarvestJob is submitted}

\item{OriginEndpointId}{[required] The ID of the OriginEndpoint that the HarvestJob will harvest from.
This cannot be changed after the HarvestJob is submitted.}

\item{S3Destination}{[required]}

\item{StartTime}{[required] The start of the time-window which will be harvested}
}
\description{
Creates a new HarvestJob record.
}
\section{Request syntax}{
\preformatted{svc$create_harvest_job(
  EndTime = "string",
  Id = "string",
  OriginEndpointId = "string",
  S3Destination = list(
    BucketName = "string",
    ManifestKey = "string",
    RoleArn = "string"
  ),
  StartTime = "string"
)
}
}

\keyword{internal}
