% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_describe_channel}
\alias{mediapackage_describe_channel}
\title{Gets details about a Channel}
\usage{
mediapackage_describe_channel(Id)
}
\arguments{
\item{Id}{[required] The ID of a Channel.}
}
\description{
Gets details about a Channel.
}
\section{Request syntax}{
\preformatted{svc$describe_channel(
  Id = "string"
)
}
}

\keyword{internal}
