% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_describe_harvest_job}
\alias{mediapackage_describe_harvest_job}
\title{Gets details about an existing HarvestJob}
\usage{
mediapackage_describe_harvest_job(Id)
}
\arguments{
\item{Id}{[required] The ID of the HarvestJob.}
}
\description{
Gets details about an existing HarvestJob.
}
\section{Request syntax}{
\preformatted{svc$describe_harvest_job(
  Id = "string"
)
}
}

\keyword{internal}
