% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_describe_origin_endpoint}
\alias{mediapackage_describe_origin_endpoint}
\title{Gets details about an existing OriginEndpoint}
\usage{
mediapackage_describe_origin_endpoint(Id)
}
\arguments{
\item{Id}{[required] The ID of the OriginEndpoint.}
}
\description{
Gets details about an existing OriginEndpoint.
}
\section{Request syntax}{
\preformatted{mediapackage$describe_origin_endpoint(
  Id = "string"
)
}
}

\keyword{internal}
