% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_list_channels}
\alias{mediapackage_list_channels}
\title{Returns a collection of Channels}
\usage{
mediapackage_list_channels(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{Upper bound on number of records to return.}

\item{NextToken}{A token used to resume pagination from the end of a previous request.}
}
\description{
Returns a collection of Channels.
}
\section{Request syntax}{
\preformatted{mediapackage$list_channels(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
