% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_update_origin_endpoint}
\alias{mediapackage_update_origin_endpoint}
\title{Updates an existing OriginEndpoint}
\usage{
mediapackage_update_origin_endpoint(CmafPackage, DashPackage,
  Description, HlsPackage, Id, ManifestName, MssPackage,
  StartoverWindowSeconds, TimeDelaySeconds, Whitelist)
}
\arguments{
\item{CmafPackage}{}

\item{DashPackage}{}

\item{Description}{A short text description of the OriginEndpoint.}

\item{HlsPackage}{}

\item{Id}{[required] The ID of the OriginEndpoint to update.}

\item{ManifestName}{A short string that will be appended to the end of the Endpoint URL.}

\item{MssPackage}{}

\item{StartoverWindowSeconds}{Maximum duration (in seconds) of content to retain for startover playback.
If not specified, startover playback will be disabled for the OriginEndpoint.}

\item{TimeDelaySeconds}{Amount of delay (in seconds) to enforce on the playback of live content.
If not specified, there will be no time delay in effect for the OriginEndpoint.}

\item{Whitelist}{A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.}
}
\description{
Updates an existing OriginEndpoint.
}
\section{Request syntax}{
\preformatted{svc$update_origin_endpoint(
  CmafPackage = list(
    Encryption = list(
      KeyRotationIntervalSeconds = 123,
      SpekeKeyProvider = list(
        CertificateArn = "string",
        ResourceId = "string",
        RoleArn = "string",
        SystemIds = list(
          "string"
        ),
        Url = "string"
      )
    ),
    HlsManifests = list(
      list(
        AdMarkers = "NONE"|"SCTE35_ENHANCED"|"PASSTHROUGH",
        AdTriggers = list(
          "SPLICE_INSERT"|"BREAK"|"PROVIDER_ADVERTISEMENT"|"DISTRIBUTOR_ADVERTISEMENT"|"PROVIDER_PLACEMENT_OPPORTUNITY"|"DISTRIBUTOR_PLACEMENT_OPPORTUNITY"|"PROVIDER_OVERLAY_PLACEMENT_OPPORTUNITY"|"DISTRIBUTOR_OVERLAY_PLACEMENT_OPPORTUNITY"
        ),
        AdsOnDeliveryRestrictions = "NONE"|"RESTRICTED"|"UNRESTRICTED"|"BOTH",
        Id = "string",
        IncludeIframeOnlyStream = TRUE|FALSE,
        ManifestName = "string",
        PlaylistType = "NONE"|"EVENT"|"VOD",
        PlaylistWindowSeconds = 123,
        ProgramDateTimeIntervalSeconds = 123
      )
    ),
    SegmentDurationSeconds = 123,
    SegmentPrefix = "string",
    StreamSelection = list(
      MaxVideoBitsPerSecond = 123,
      MinVideoBitsPerSecond = 123,
      StreamOrder = "ORIGINAL"|"VIDEO_BITRATE_ASCENDING"|"VIDEO_BITRATE_DESCENDING"
    )
  ),
  DashPackage = list(
    AdTriggers = list(
      "SPLICE_INSERT"|"BREAK"|"PROVIDER_ADVERTISEMENT"|"DISTRIBUTOR_ADVERTISEMENT"|"PROVIDER_PLACEMENT_OPPORTUNITY"|"DISTRIBUTOR_PLACEMENT_OPPORTUNITY"|"PROVIDER_OVERLAY_PLACEMENT_OPPORTUNITY"|"DISTRIBUTOR_OVERLAY_PLACEMENT_OPPORTUNITY"
    ),
    AdsOnDeliveryRestrictions = "NONE"|"RESTRICTED"|"UNRESTRICTED"|"BOTH",
    Encryption = list(
      KeyRotationIntervalSeconds = 123,
      SpekeKeyProvider = list(
        CertificateArn = "string",
        ResourceId = "string",
        RoleArn = "string",
        SystemIds = list(
          "string"
        ),
        Url = "string"
      )
    ),
    ManifestLayout = "FULL"|"COMPACT",
    ManifestWindowSeconds = 123,
    MinBufferTimeSeconds = 123,
    MinUpdatePeriodSeconds = 123,
    PeriodTriggers = list(
      "ADS"
    ),
    Profile = "NONE"|"HBBTV_1_5",
    SegmentDurationSeconds = 123,
    SegmentTemplateFormat = "NUMBER_WITH_TIMELINE"|"TIME_WITH_TIMELINE"|"NUMBER_WITH_DURATION",
    StreamSelection = list(
      MaxVideoBitsPerSecond = 123,
      MinVideoBitsPerSecond = 123,
      StreamOrder = "ORIGINAL"|"VIDEO_BITRATE_ASCENDING"|"VIDEO_BITRATE_DESCENDING"
    ),
    SuggestedPresentationDelaySeconds = 123
  ),
  Description = "string",
  HlsPackage = list(
    AdMarkers = "NONE"|"SCTE35_ENHANCED"|"PASSTHROUGH",
    AdTriggers = list(
      "SPLICE_INSERT"|"BREAK"|"PROVIDER_ADVERTISEMENT"|"DISTRIBUTOR_ADVERTISEMENT"|"PROVIDER_PLACEMENT_OPPORTUNITY"|"DISTRIBUTOR_PLACEMENT_OPPORTUNITY"|"PROVIDER_OVERLAY_PLACEMENT_OPPORTUNITY"|"DISTRIBUTOR_OVERLAY_PLACEMENT_OPPORTUNITY"
    ),
    AdsOnDeliveryRestrictions = "NONE"|"RESTRICTED"|"UNRESTRICTED"|"BOTH",
    Encryption = list(
      ConstantInitializationVector = "string",
      EncryptionMethod = "AES_128"|"SAMPLE_AES",
      KeyRotationIntervalSeconds = 123,
      RepeatExtXKey = TRUE|FALSE,
      SpekeKeyProvider = list(
        CertificateArn = "string",
        ResourceId = "string",
        RoleArn = "string",
        SystemIds = list(
          "string"
        ),
        Url = "string"
      )
    ),
    IncludeIframeOnlyStream = TRUE|FALSE,
    PlaylistType = "NONE"|"EVENT"|"VOD",
    PlaylistWindowSeconds = 123,
    ProgramDateTimeIntervalSeconds = 123,
    SegmentDurationSeconds = 123,
    StreamSelection = list(
      MaxVideoBitsPerSecond = 123,
      MinVideoBitsPerSecond = 123,
      StreamOrder = "ORIGINAL"|"VIDEO_BITRATE_ASCENDING"|"VIDEO_BITRATE_DESCENDING"
    ),
    UseAudioRenditionGroup = TRUE|FALSE
  ),
  Id = "string",
  ManifestName = "string",
  MssPackage = list(
    Encryption = list(
      SpekeKeyProvider = list(
        CertificateArn = "string",
        ResourceId = "string",
        RoleArn = "string",
        SystemIds = list(
          "string"
        ),
        Url = "string"
      )
    ),
    ManifestWindowSeconds = 123,
    SegmentDurationSeconds = 123,
    StreamSelection = list(
      MaxVideoBitsPerSecond = 123,
      MinVideoBitsPerSecond = 123,
      StreamOrder = "ORIGINAL"|"VIDEO_BITRATE_ASCENDING"|"VIDEO_BITRATE_DESCENDING"
    )
  ),
  StartoverWindowSeconds = 123,
  TimeDelaySeconds = 123,
  Whitelist = list(
    "string"
  )
)
}
}

\keyword{internal}
