% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_service.R
\name{mediapackagevod}
\alias{mediapackagevod}
\title{AWS Elemental MediaPackage VOD}
\usage{
mediapackagevod(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
AWS Elemental MediaPackage VOD
}
\section{Service syntax}{
\preformatted{svc <- mediapackagevod(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=mediapackagevod_create_asset]{create_asset} \tab Creates a new MediaPackage VOD Asset resource \cr
\link[=mediapackagevod_create_packaging_configuration]{create_packaging_configuration} \tab Creates a new MediaPackage VOD PackagingConfiguration resource \cr
\link[=mediapackagevod_create_packaging_group]{create_packaging_group} \tab Creates a new MediaPackage VOD PackagingGroup resource \cr
\link[=mediapackagevod_delete_asset]{delete_asset} \tab Deletes an existing MediaPackage VOD Asset resource \cr
\link[=mediapackagevod_delete_packaging_configuration]{delete_packaging_configuration} \tab Deletes a MediaPackage VOD PackagingConfiguration resource \cr
\link[=mediapackagevod_delete_packaging_group]{delete_packaging_group} \tab Deletes a MediaPackage VOD PackagingGroup resource \cr
\link[=mediapackagevod_describe_asset]{describe_asset} \tab Returns a description of a MediaPackage VOD Asset resource \cr
\link[=mediapackagevod_describe_packaging_configuration]{describe_packaging_configuration} \tab Returns a description of a MediaPackage VOD PackagingConfiguration resource\cr
\link[=mediapackagevod_describe_packaging_group]{describe_packaging_group} \tab Returns a description of a MediaPackage VOD PackagingGroup resource \cr
\link[=mediapackagevod_list_assets]{list_assets} \tab Returns a collection of MediaPackage VOD Asset resources \cr
\link[=mediapackagevod_list_packaging_configurations]{list_packaging_configurations} \tab Returns a collection of MediaPackage VOD PackagingConfiguration resources \cr
\link[=mediapackagevod_list_packaging_groups]{list_packaging_groups} \tab Returns a collection of MediaPackage VOD PackagingGroup resources
}
}

\examples{
\donttest{svc <- mediapackagevod()
svc$create_asset(
  Foo = 123
)}

}
