% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_create_asset}
\alias{mediapackagevod_create_asset}
\title{Creates a new MediaPackage VOD Asset resource}
\usage{
mediapackagevod_create_asset(Id, PackagingGroupId, ResourceId,
  SourceArn, SourceRoleArn)
}
\arguments{
\item{Id}{[required] The unique identifier for the Asset.}

\item{PackagingGroupId}{[required] The ID of the PackagingGroup for the Asset.}

\item{ResourceId}{The resource ID to include in SPEKE key requests.}

\item{SourceArn}{[required] ARN of the source object in S3.}

\item{SourceRoleArn}{[required] The IAM role ARN used to access the source S3 bucket.}
}
\description{
Creates a new MediaPackage VOD Asset resource.
}
\section{Request syntax}{
\preformatted{svc$create_asset(
  Id = "string",
  PackagingGroupId = "string",
  ResourceId = "string",
  SourceArn = "string",
  SourceRoleArn = "string"
)
}
}

\keyword{internal}
