% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_delete_asset}
\alias{mediapackagevod_delete_asset}
\title{Deletes an existing MediaPackage VOD Asset resource}
\usage{
mediapackagevod_delete_asset(Id)
}
\arguments{
\item{Id}{[required] The ID of the MediaPackage VOD Asset resource to delete.}
}
\description{
Deletes an existing MediaPackage VOD Asset resource.
}
\section{Request syntax}{
\preformatted{svc$delete_asset(
  Id = "string"
)
}
}

\keyword{internal}
