% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_delete_packaging_configuration}
\alias{mediapackagevod_delete_packaging_configuration}
\title{Deletes a MediaPackage VOD PackagingConfiguration resource}
\usage{
mediapackagevod_delete_packaging_configuration(Id)
}
\arguments{
\item{Id}{[required] The ID of the MediaPackage VOD PackagingConfiguration resource to delete.}
}
\description{
Deletes a MediaPackage VOD PackagingConfiguration resource.
}
\section{Request syntax}{
\preformatted{svc$delete_packaging_configuration(
  Id = "string"
)
}
}

\keyword{internal}
