% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_delete_packaging_group}
\alias{mediapackagevod_delete_packaging_group}
\title{Deletes a MediaPackage VOD PackagingGroup resource}
\usage{
mediapackagevod_delete_packaging_group(Id)
}
\arguments{
\item{Id}{[required] The ID of the MediaPackage VOD PackagingGroup resource to delete.}
}
\description{
Deletes a MediaPackage VOD PackagingGroup resource.
}
\section{Request syntax}{
\preformatted{svc$delete_packaging_group(
  Id = "string"
)
}
}

\keyword{internal}
