% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_describe_asset}
\alias{mediapackagevod_describe_asset}
\title{Returns a description of a MediaPackage VOD Asset resource}
\usage{
mediapackagevod_describe_asset(Id)
}
\arguments{
\item{Id}{[required] The ID of an MediaPackage VOD Asset resource.}
}
\description{
Returns a description of a MediaPackage VOD Asset resource.
}
\section{Request syntax}{
\preformatted{svc$describe_asset(
  Id = "string"
)
}
}

\keyword{internal}
