% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_describe_packaging_configuration}
\alias{mediapackagevod_describe_packaging_configuration}
\title{Returns a description of a MediaPackage VOD PackagingConfiguration
resource}
\usage{
mediapackagevod_describe_packaging_configuration(Id)
}
\arguments{
\item{Id}{[required] The ID of a MediaPackage VOD PackagingConfiguration resource.}
}
\description{
Returns a description of a MediaPackage VOD PackagingConfiguration resource.
}
\section{Request syntax}{
\preformatted{svc$describe_packaging_configuration(
  Id = "string"
)
}
}

\keyword{internal}
