% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_describe_packaging_group}
\alias{mediapackagevod_describe_packaging_group}
\title{Returns a description of a MediaPackage VOD PackagingGroup resource}
\usage{
mediapackagevod_describe_packaging_group(Id)
}
\arguments{
\item{Id}{[required] The ID of a MediaPackage VOD PackagingGroup resource.}
}
\description{
Returns a description of a MediaPackage VOD PackagingGroup resource.
}
\section{Request syntax}{
\preformatted{svc$describe_packaging_group(
  Id = "string"
)
}
}

\keyword{internal}
