% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_list_assets}
\alias{mediapackagevod_list_assets}
\title{Returns a collection of MediaPackage VOD Asset resources}
\usage{
mediapackagevod_list_assets(MaxResults, NextToken, PackagingGroupId)
}
\arguments{
\item{MaxResults}{Upper bound on number of records to return.}

\item{NextToken}{A token used to resume pagination from the end of a previous request.}

\item{PackagingGroupId}{Returns Assets associated with the specified PackagingGroup.}
}
\description{
Returns a collection of MediaPackage VOD Asset resources.
}
\section{Request syntax}{
\preformatted{svc$list_assets(
  MaxResults = 123,
  NextToken = "string",
  PackagingGroupId = "string"
)
}
}

\keyword{internal}
