% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_list_packaging_configurations}
\alias{mediapackagevod_list_packaging_configurations}
\title{Returns a collection of MediaPackage VOD PackagingConfiguration
resources}
\usage{
mediapackagevod_list_packaging_configurations(MaxResults, NextToken,
  PackagingGroupId)
}
\arguments{
\item{MaxResults}{Upper bound on number of records to return.}

\item{NextToken}{A token used to resume pagination from the end of a previous request.}

\item{PackagingGroupId}{Returns MediaPackage VOD PackagingConfigurations associated with the specified PackagingGroup.}
}
\description{
Returns a collection of MediaPackage VOD PackagingConfiguration resources.
}
\section{Request syntax}{
\preformatted{svc$list_packaging_configurations(
  MaxResults = 123,
  NextToken = "string",
  PackagingGroupId = "string"
)
}
}

\keyword{internal}
