% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_list_packaging_groups}
\alias{mediapackagevod_list_packaging_groups}
\title{Returns a collection of MediaPackage VOD PackagingGroup resources}
\usage{
mediapackagevod_list_packaging_groups(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{Upper bound on number of records to return.}

\item{NextToken}{A token used to resume pagination from the end of a previous request.}
}
\description{
Returns a collection of MediaPackage VOD PackagingGroup resources.
}
\section{Request syntax}{
\preformatted{svc$list_packaging_groups(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
