% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_list_tags_for_resource}
\alias{mediapackagevod_list_tags_for_resource}
\title{Returns a list of the tags assigned to the specified resource}
\usage{
mediapackagevod_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.}
}
\description{
Returns a list of the tags assigned to the specified resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
