% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_tag_resource}
\alias{mediapackagevod_tag_resource}
\title{Adds tags to the specified resource}
\usage{
mediapackagevod_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.}

\item{Tags}{[required] A collection of tags associated with a resource}
}
\description{
Adds tags to the specified resource. You can specify one or more tags to add.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
