% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_untag_resource}
\alias{mediapackagevod_untag_resource}
\title{Removes tags from the specified resource}
\usage{
mediapackagevod_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.}

\item{TagKeys}{[required] A comma-separated list of the tag keys to remove from the resource.}
}
\description{
Removes tags from the specified resource. You can specify one or more tags to remove.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
