% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_update_packaging_group}
\alias{mediapackagevod_update_packaging_group}
\title{Updates a specific packaging group}
\usage{
mediapackagevod_update_packaging_group(Authorization, Id)
}
\arguments{
\item{Authorization}{}

\item{Id}{[required] The ID of a MediaPackage VOD PackagingGroup resource.}
}
\description{
Updates a specific packaging group. You can't change the id attribute or any other system-generated attributes.
}
\section{Request syntax}{
\preformatted{svc$update_packaging_group(
  Authorization = list(
    CdnIdentifierSecret = "string",
    SecretsRoleArn = "string"
  ),
  Id = "string"
)
}
}

\keyword{internal}
