% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_create_container}
\alias{mediastore_create_container}
\title{Creates a storage container to hold objects}
\usage{
mediastore_create_container(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name for the container. The name must be from 1 to 255 characters.
Container names must be unique to your AWS account within a specific
region. As an example, you could create a container named \code{movies} in
every region, as long as you don't have an existing container with that
name.}
}
\description{
Creates a storage container to hold objects. A container is similar to a
bucket in the Amazon S3 service.
}
\section{Request syntax}{
\preformatted{svc$create_container(
  ContainerName = "string"
)
}
}

\keyword{internal}
