% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_delete_lifecycle_policy}
\alias{mediastore_delete_lifecycle_policy}
\title{Removes an object lifecycle policy from a container}
\usage{
mediastore_delete_lifecycle_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that holds the object lifecycle policy.}
}
\description{
Removes an object lifecycle policy from a container.
}
\section{Request syntax}{
\preformatted{mediastore$delete_lifecycle_policy(
  ContainerName = "string"
)
}
}

\keyword{internal}
