% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_get_cors_policy}
\alias{mediastore_get_cors_policy}
\title{Returns the cross-origin resource sharing (CORS) configuration
information that is set for the container}
\usage{
mediastore_get_cors_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that the policy is assigned to.}
}
\description{
Returns the cross-origin resource sharing (CORS) configuration
information that is set for the container.
}
\details{
To use this operation, you must have permission to perform the
\code{MediaStore:GetCorsPolicy} action. By default, the container owner has
this permission and can grant it to others.
}
\section{Request syntax}{
\preformatted{mediastore$get_cors_policy(
  ContainerName = "string"
)
}
}

\keyword{internal}
