% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_get_lifecycle_policy}
\alias{mediastore_get_lifecycle_policy}
\title{Retrieves the object lifecycle policy that is assigned to a container}
\usage{
mediastore_get_lifecycle_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that the object lifecycle policy is assigned
to.}
}
\description{
Retrieves the object lifecycle policy that is assigned to a container.
}
\section{Request syntax}{
\preformatted{mediastore$get_lifecycle_policy(
  ContainerName = "string"
)
}
}

\keyword{internal}
