% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_list_tags_for_resource}
\alias{mediastore_list_tags_for_resource}
\title{Returns a list of the tags assigned to the specified container}
\usage{
mediastore_list_tags_for_resource(Resource)
}
\arguments{
\item{Resource}{[required] The Amazon Resource Name (ARN) for the container.}
}
\description{
Returns a list of the tags assigned to the specified container.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  Resource = "string"
)
}
}

\keyword{internal}
