% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_put_container_policy}
\alias{mediastore_put_container_policy}
\title{Creates an access policy for the specified container to restrict the
users and clients that can access it}
\usage{
mediastore_put_container_policy(ContainerName, Policy)
}
\arguments{
\item{ContainerName}{[required] The name of the container.}

\item{Policy}{[required] The contents of the policy, which includes the following:
\itemize{
\item One \code{Version} tag
\item One \code{Statement} tag that contains the standard tags for the policy.
}}
}
\description{
Creates an access policy for the specified container to restrict the
users and clients that can access it. For information about the data
that is included in an access policy, see the \href{https://docs.aws.amazon.com/iam/index.html}{AWS Identity and Access Management User Guide}.

For this release of the REST API, you can create only one policy for a
container. If you enter
\code{\link[=mediastore_put_container_policy]{put_container_policy}} twice, the
second command modifies the existing policy.
}
\section{Request syntax}{
\preformatted{svc$put_container_policy(
  ContainerName = "string",
  Policy = "string"
)
}
}

\keyword{internal}
