% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_start_access_logging}
\alias{mediastore_start_access_logging}
\title{Starts access logging on the specified container}
\usage{
mediastore_start_access_logging(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that you want to start access logging on.}
}
\description{
Starts access logging on the specified container. When you enable access
logging on a container, MediaStore delivers access logs for objects
stored in that container to Amazon CloudWatch Logs.
}
\section{Request syntax}{
\preformatted{svc$start_access_logging(
  ContainerName = "string"
)
}
}

\keyword{internal}
