% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_stop_access_logging}
\alias{mediastore_stop_access_logging}
\title{Stops access logging on the specified container}
\usage{
mediastore_stop_access_logging(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that you want to stop access logging on.}
}
\description{
Stops access logging on the specified container. When you stop access
logging on a container, MediaStore stops sending access logs to Amazon
CloudWatch Logs. These access logs are not saved and are not
retrievable.
}
\section{Request syntax}{
\preformatted{svc$stop_access_logging(
  ContainerName = "string"
)
}
}

\keyword{internal}
