% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_untag_resource}
\alias{mediastore_untag_resource}
\title{Removes tags from the specified container}
\usage{
mediastore_untag_resource(Resource, TagKeys)
}
\arguments{
\item{Resource}{[required] The Amazon Resource Name (ARN) for the container.}

\item{TagKeys}{[required] A comma-separated list of keys for tags that you want to remove from the
container. For example, if your container has two tags
(customer:CompanyA and priority:High) and you want to remove one of the
tags (priority:High), you specify the key for the tag that you want to
remove (priority).}
}
\description{
Removes tags from the specified container. You can specify one or more
tags to remove.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  Resource = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
