% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{mediastoredata}
\alias{mediastoredata}
\title{AWS Elemental MediaStore Data Plane}
\usage{
mediastoredata()
}
\description{
An AWS Elemental MediaStore asset is an object, similar to an object in
the Amazon S3 service. Objects are the fundamental entities that are
stored in AWS Elemental MediaStore.
}
\section{Operations}{

\tabular{ll}{
\link[=mediastoredata_delete_object]{delete_object} \tab Deletes an object at the specified path \cr
\link[=mediastoredata_describe_object]{describe_object} \tab Gets the headers for an object at the specified path \cr
\link[=mediastoredata_get_object]{get_object} \tab Downloads the object at the specified path \cr
\link[=mediastoredata_list_items]{list_items} \tab Provides a list of metadata entries about folders and objects in the specified folder\cr
\link[=mediastoredata_put_object]{put_object} \tab Uploads an object to the specified path
}
}

\examples{
\donttest{svc <- mediastoredata()
svc$delete_object(
  Foo = 123
)}

}
