% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastoredata_operations.R
\name{mediastoredata_describe_object}
\alias{mediastoredata_describe_object}
\title{Gets the headers for an object at the specified path}
\usage{
mediastoredata_describe_object(Path)
}
\arguments{
\item{Path}{[required] The path (including the file name) where the object is stored in the
container. Format: \<folder name\>/\<folder name\>/\<file name\>}
}
\description{
Gets the headers for an object at the specified path.
}
\section{Request syntax}{
\preformatted{mediastoredata$describe_object(
  Path = "string"
)
}
}

\keyword{internal}
