% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_put_playback_configuration}
\alias{mediatailor_put_playback_configuration}
\title{Adds a new configuration to AWS Elemental MediaTailor}
\usage{
mediatailor_put_playback_configuration(AdDecisionServerUrl,
  CdnConfiguration, DashConfiguration, Name, SlateAdUrl,
  TranscodeProfileName, VideoContentSourceUrl)
}
\arguments{
\item{AdDecisionServerUrl}{The URL for the ad decision server (ADS). This includes the
specification of static parameters and placeholders for dynamic
parameters. AWS Elemental MediaTailor substitutes player-specific and
session-specific parameters as needed when calling the ADS. Alternately,
for testing you can provide a static VAST URL. The maximum length is
25000 characters.}

\item{CdnConfiguration}{The configuration for using a content delivery network (CDN), like
Amazon CloudFront, for content and ad segment management.}

\item{DashConfiguration}{The configuration object for DASH content.}

\item{Name}{The identifier for the configuration.}

\item{SlateAdUrl}{The URL for a high-quality video asset to transcode and use to fill in
time that's not used by ads. AWS Elemental MediaTailor shows the slate
to fill in gaps in media content. Configuring the slate is optional for
non-VPAID configurations. For VPAID, the slate is required because AWS
Elemental MediaTailor provides it in the slots that are designated for
dynamic ad content. The slate must be a high-quality asset that contains
both audio and video.}

\item{TranscodeProfileName}{Associate this playbackConfiguration with a custom transcode profile,
overriding MediaTailor's dynamic transcoding defaults. Do not include
this field if you have not setup custom profiles with the MediaTailor
service team.}

\item{VideoContentSourceUrl}{The URL prefix for the master playlist for the stream, minus the asset
ID. The maximum length is 512 characters.}
}
\description{
Adds a new configuration to AWS Elemental MediaTailor.
}
\section{Request syntax}{
\preformatted{mediatailor$put_playback_configuration(
  AdDecisionServerUrl = "string",
  CdnConfiguration = list(
    AdSegmentUrlPrefix = "string",
    ContentSegmentUrlPrefix = "string"
  ),
  DashConfiguration = list(
    MpdLocation = "string"
  ),
  Name = "string",
  SlateAdUrl = "string",
  TranscodeProfileName = "string",
  VideoContentSourceUrl = "string"
)
}
}

\keyword{internal}
