% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_associate_discovered_resource}
\alias{migrationhub_associate_discovered_resource}
\title{Associates a discovered resource ID from Application Discovery Service
with a migration task}
\usage{
migrationhub_associate_discovered_resource(ProgressUpdateStream,
  MigrationTaskName, DiscoveredResource, DryRun)
}
\arguments{
\item{ProgressUpdateStream}{[required] The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{[required] The identifier given to the MigrationTask. \emph{Do not store personal data
in this field.}}

\item{DiscoveredResource}{[required] Object representing a Resource.}

\item{DryRun}{Optional boolean flag to indicate whether any effect should take place.
Used to test if the caller has permission to make the call.}
}
\description{
Associates a discovered resource ID from Application Discovery Service
with a migration task.
}
\section{Request syntax}{
\preformatted{svc$associate_discovered_resource(
  ProgressUpdateStream = "string",
  MigrationTaskName = "string",
  DiscoveredResource = list(
    ConfigurationId = "string",
    Description = "string"
  ),
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
