% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_describe_migration_task}
\alias{migrationhub_describe_migration_task}
\title{Retrieves a list of all attributes associated with a specific migration
task}
\usage{
migrationhub_describe_migration_task(ProgressUpdateStream,
  MigrationTaskName)
}
\arguments{
\item{ProgressUpdateStream}{[required] The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{[required] The identifier given to the MigrationTask. \emph{Do not store personal data
in this field.}}
}
\description{
Retrieves a list of all attributes associated with a specific migration
task.
}
\section{Request syntax}{
\preformatted{svc$describe_migration_task(
  ProgressUpdateStream = "string",
  MigrationTaskName = "string"
)
}
}

\keyword{internal}
