% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhubconfig_service.R
\name{migrationhubconfig}
\alias{migrationhubconfig}
\title{AWS Migration Hub Config}
\usage{
migrationhubconfig(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
The AWS Migration Hub home region APIs are available specifically for
working with your Migration Hub home region. You can use these APIs to
determine a home region, as well as to create and work with controls
that describe the home region.
\itemize{
\item You must make API calls for write actions (create, notify,
associate, disassociate, import, or put) while in your home region,
or a \code{HomeRegionNotSetException} error is returned.
\item API calls for read actions (list, describe, stop, and delete) are
permitted outside of your home region.
\item If you call a write API outside the home region, an
\code{InvalidInputException} is returned.
\item You can call \code{GetHomeRegion} action to obtain the account\'s
Migration Hub home region.
}

For specific API usage, see the sections that follow in this AWS
Migration Hub Home Region API reference.
}
\section{Service syntax}{
\preformatted{svc <- migrationhubconfig(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=migrationhubconfig_create_home_region_control]{create_home_region_control} \tab This API sets up the home region for the calling account only \cr
\link[=migrationhubconfig_describe_home_region_controls]{describe_home_region_controls} \tab This API permits filtering on the ControlId and HomeRegion fields\cr
\link[=migrationhubconfig_get_home_region]{get_home_region} \tab Returns the calling account’s home region, if configured
}
}

\examples{
\dontrun{
svc <- migrationhubconfig()
svc$create_home_region_control(
  Foo = 123
)
}

}
