% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhubconfig_operations.R
\name{migrationhubconfig_create_home_region_control}
\alias{migrationhubconfig_create_home_region_control}
\title{This API sets up the home region for the calling account only}
\usage{
migrationhubconfig_create_home_region_control(HomeRegion, Target,
  DryRun)
}
\arguments{
\item{HomeRegion}{[required] The name of the home region of the calling account.}

\item{Target}{[required] The account for which this command sets up a home region control. The
\code{Target} is always of type \code{ACCOUNT}.}

\item{DryRun}{Optional Boolean flag to indicate whether any effect should take place.
It tests whether the caller has permission to make the call.}
}
\description{
This API sets up the home region for the calling account only.
}
\section{Request syntax}{
\preformatted{svc$create_home_region_control(
  HomeRegion = "string",
  Target = list(
    Type = "ACCOUNT",
    Id = "string"
  ),
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
