% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhubconfig_operations.R
\name{migrationhubconfig_describe_home_region_controls}
\alias{migrationhubconfig_describe_home_region_controls}
\title{This API permits filtering on the ControlId and HomeRegion fields}
\usage{
migrationhubconfig_describe_home_region_controls(ControlId, HomeRegion,
  Target, MaxResults, NextToken)
}
\arguments{
\item{ControlId}{The \code{ControlID} is a unique identifier string of your
\code{HomeRegionControl} object.}

\item{HomeRegion}{The name of the home region you\'d like to view.}

\item{Target}{The target parameter specifies the identifier to which the home region
is applied, which is always of type \code{ACCOUNT}. It applies the home
region to the current \code{ACCOUNT}.}

\item{MaxResults}{The maximum number of filtering results to display per page.}

\item{NextToken}{If a \code{NextToken} was returned by a previous call, more results are
available. To retrieve the next page of results, make the call again
using the returned token in \code{NextToken}.}
}
\description{
This API permits filtering on the \code{ControlId} and \code{HomeRegion} fields.
}
\section{Request syntax}{
\preformatted{svc$describe_home_region_controls(
  ControlId = "string",
  HomeRegion = "string",
  Target = list(
    Type = "ACCOUNT",
    Id = "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
