% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhubconfig_operations.R
\name{migrationhubconfig_get_home_region}
\alias{migrationhubconfig_get_home_region}
\title{Returns the calling account’s home region, if configured}
\usage{
migrationhubconfig_get_home_region()
}
\description{
Returns the calling account’s home region, if configured. This API is
used by other AWS services to determine the regional endpoint for
calling AWS Application Discovery Service and Migration Hub. You must
call \code{GetHomeRegion} at least once before you call any other AWS
Application Discovery Service and AWS Migration Hub APIs, to obtain the
account's Migration Hub home region.
}
\section{Request syntax}{
\preformatted{svc$get_home_region()
}
}

\keyword{internal}
