% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filter_plots.R
\name{mini_paws}
\alias{mini_paws}
\title{Run PAWS analysis for a single CSV (rather than in batch)}
\usage{
mini_paws(
  csv_or_path,
  manual_scale_factor = NA,
  p_cutoff = 0,
  filter = "none",
  body_part = "center",
  reference_distance,
  fps = 2000,
  savgol_window_length = 11,
  median_window_length = 11,
  average_window_length = 11,
  shake_threshold = 0.35,
  window_threshold = 0.5
)
}
\arguments{
\item{csv_or_path}{Either a csv loaded as an R object, or the full path to that CSV.}

\item{manual_scale_factor}{The 'real-world-length' to pixel conversion factor (i.e. millimeters/pixel).
If you are using two reference points, you can ignore this parameter.}

\item{p_cutoff}{The confidence value at which all tracked points below this
threshold are replaced by linear interpolation between two higher-confidence points.}

\item{filter}{The filter chosen to smooth tracked trajectories.
Options are "none", "savitzky-golay", "median", or "average" (average recommended)}

\item{body_part}{The body_parts you wish to plot.}

\item{reference_distance}{The 'real-world-length' between your reference objects (can be in mm, cm, etc).
If you indicate a \code{manual_scale_factor}, this setting will be overrided by that factor.}

\item{fps}{The frames per second of your CSV.}

\item{savgol_window_length}{The rolling window length of Savitzky-Golay filter smoothing to apply
to your tracking trajectory.}

\item{median_window_length}{The rolling window length of median filter smoothing to apply
to your tracking trajectory.}

\item{average_window_length}{The rolling window length of average filter smoothing to apply
to your tracking trajectory.}

\item{shake_threshold}{The threshold to examine a given window of tracking activity
for shaking behaviors. Higher values apply a more conservative shaking threshold (Tip:
if shake segmentation does not match what you see in a video, you can fine-tune the threshold
to match your observations).}

\item{window_threshold}{The threshold to examine a given window of tracking activity
for a withdrawal behavior. Higher values apply a more conservative window threshold.}
}
\value{
A data frame in RStudio containing PAWS scores for a single file.
}
\description{
This function loads a single CSV containing tracking data. The function then scores the tracked
data and outputs PAWS scores.
}
