% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_service.R
\name{mobile}
\alias{mobile}
\title{AWS Mobile}
\usage{
mobile(config = NULL)
}
\arguments{
\item{config}{An optional list of custom configurations for the service. Currently
supports adding custom credentials, endpoint, and region.}
}
\description{
AWS Mobile Service provides mobile app and website developers with
capabilities required to configure AWS resources and bootstrap their
developer desktop projects with the necessary SDKs, constants, tools and
samples to make use of those resources.
}
\section{Service syntax}{
\preformatted{svc <- mobile(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string",
        provider_name = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=mobile_create_project]{create_project} \tab Creates an AWS Mobile Hub project \cr
\link[=mobile_delete_project]{delete_project} \tab Delets a project in AWS Mobile Hub \cr
\link[=mobile_describe_bundle]{describe_bundle} \tab Get the bundle details for the requested bundle id \cr
\link[=mobile_describe_project]{describe_project} \tab Gets details about a project in AWS Mobile Hub \cr
\link[=mobile_export_bundle]{export_bundle} \tab Generates customized software development kit (SDK) and or tool packages used to integrate mobile web or mobile app clients with backend AWS resources\cr
\link[=mobile_export_project]{export_project} \tab Exports project configuration to a snapshot which can be downloaded and shared \cr
\link[=mobile_list_bundles]{list_bundles} \tab List all available bundles \cr
\link[=mobile_list_projects]{list_projects} \tab Lists projects in AWS Mobile Hub \cr
\link[=mobile_update_project]{update_project} \tab Update an existing project
}
}

\examples{
\donttest{svc <- mobile()
svc$create_project(
  Foo = 123
)}

}
