% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_operations.R
\name{mobile_delete_project}
\alias{mobile_delete_project}
\title{Delets a project in AWS Mobile Hub}
\usage{
mobile_delete_project(projectId)
}
\arguments{
\item{projectId}{[required] Unique project identifier.}
}
\description{
Delets a project in AWS Mobile Hub.
}
\section{Request syntax}{
\preformatted{mobile$delete_project(
  projectId = "string"
)
}
}

\keyword{internal}
