% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_operations.R
\name{mobile_list_projects}
\alias{mobile_list_projects}
\title{Lists projects in AWS Mobile Hub}
\usage{
mobile_list_projects(maxResults, nextToken)
}
\arguments{
\item{maxResults}{Maximum number of records to list in a single response.}

\item{nextToken}{Pagination token. Set to null to start listing projects from start. If
non-null pagination token is returned in a result, then pass its value
in here in another request to list more projects.}
}
\description{
Lists projects in AWS Mobile Hub.
}
\section{Request syntax}{
\preformatted{svc$list_projects(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
