% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_delete_tags}
\alias{mq_delete_tags}
\title{Remove a tag from a resource}
\usage{
mq_delete_tags(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] the Amazon Resource Name (ARN)}

\item{TagKeys}{[required] An array of tag keys to delete}
}
\description{
Remove a tag from a resource.
}
\section{Request syntax}{
\preformatted{mq$delete_tags(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
