% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_create_worker_block}
\alias{mturk_create_worker_block}
\title{The CreateWorkerBlock operation allows you to prevent a Worker from
working on your HITs}
\usage{
mturk_create_worker_block(WorkerId, Reason)
}
\arguments{
\item{WorkerId}{[required] The ID of the Worker to block.}

\item{Reason}{[required] A message explaining the reason for blocking the Worker. This parameter
enables you to keep track of your Workers. The Worker does not see this
message.}
}
\description{
The \code{CreateWorkerBlock} operation allows you to prevent a Worker from
working on your HITs. For example, you can block a Worker who is
producing poor quality work. You can block up to 100,000 Workers.
}
\section{Request syntax}{
\preformatted{mturk$create_worker_block(
  WorkerId = "string",
  Reason = "string"
)
}
}

\keyword{internal}
