% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_disassociate_qualification_from_worker}
\alias{mturk_disassociate_qualification_from_worker}
\title{The DisassociateQualificationFromWorker revokes a previously granted
Qualification from a user}
\usage{
mturk_disassociate_qualification_from_worker(WorkerId,
  QualificationTypeId, Reason)
}
\arguments{
\item{WorkerId}{[required] The ID of the Worker who possesses the Qualification to be revoked.}

\item{QualificationTypeId}{[required] The ID of the Qualification type of the Qualification to be revoked.}

\item{Reason}{A text message that explains why the Qualification was revoked. The user
who had the Qualification sees this message.}
}
\description{
The
\code{\link[=mturk_disassociate_qualification_from_worker]{disassociate_qualification_from_worker}}
revokes a previously granted Qualification from a user.

You can provide a text message explaining why the Qualification was
revoked. The user who had the Qualification can see this message.
}
\section{Request syntax}{
\preformatted{svc$disassociate_qualification_from_worker(
  WorkerId = "string",
  QualificationTypeId = "string",
  Reason = "string"
)
}
}

\keyword{internal}
