% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_get_hit}
\alias{mturk_get_hit}
\title{The GetHIT operation retrieves the details of the specified HIT}
\usage{
mturk_get_hit(HITId)
}
\arguments{
\item{HITId}{[required] The ID of the HIT to be retrieved.}
}
\description{
The \code{\link[=mturk_get_hit]{get_hit}} operation retrieves the details of the
specified HIT.
}
\section{Request syntax}{
\preformatted{svc$get_hit(
  HITId = "string"
)
}
}

\keyword{internal}
