% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_create_cli_token}
\alias{mwaa_create_cli_token}
\title{Create a CLI token to use Airflow CLI}
\usage{
mwaa_create_cli_token(Name)
}
\arguments{
\item{Name}{[required] Create a CLI token request for a MWAA environment.}
}
\description{
Create a CLI token to use Airflow CLI.
}
\section{Request syntax}{
\preformatted{svc$create_cli_token(
  Name = "string"
)
}
}

\keyword{internal}
