% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_create_environment}
\alias{mwaa_create_environment}
\title{JSON blob that describes the environment to create}
\usage{
mwaa_create_environment(AirflowConfigurationOptions, AirflowVersion,
  DagS3Path, EnvironmentClass, ExecutionRoleArn, KmsKey,
  LoggingConfiguration, MaxWorkers, Name, NetworkConfiguration,
  PluginsS3ObjectVersion, PluginsS3Path, RequirementsS3ObjectVersion,
  RequirementsS3Path, SourceBucketArn, Tags, WebserverAccessMode,
  WeeklyMaintenanceWindowStart)
}
\arguments{
\item{AirflowConfigurationOptions}{The Apache Airflow configuration setting you want to override in your
environment. For more information, see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/configuring-env-variables.html}{Environment configuration}.}

\item{AirflowVersion}{The Apache Airflow version you want to use for your environment.}

\item{DagS3Path}{[required] The relative path to the DAG folder on your Amazon S3 storage bucket.
For example, \code{dags}. For more information, see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/}{Importing DAGs on Amazon MWAA}.}

\item{EnvironmentClass}{The environment class you want to use for your environment. The
environment class determines the size of the containers and database
used for your Apache Airflow services.}

\item{ExecutionRoleArn}{[required] The Amazon Resource Name (ARN) of the execution role for your
environment. An execution role is an AWS Identity and Access Management
(IAM) role that grants MWAA permission to access AWS services and
resources used by your environment. For example,
\verb{arn:aws:iam::123456789:role/my-execution-role}. For more information,
see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/manage-access.html}{Managing access to Amazon Managed Workflows for Apache Airflow}.}

\item{KmsKey}{The AWS Key Management Service (KMS) key to encrypt and decrypt the data
in your environment. You can use an AWS KMS key managed by MWAA, or a
custom KMS key (advanced). For more information, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/concepts.html?icmpid=docs_console_unmapped#master_keys}{Customer master keys (CMKs)}
in the AWS KMS developer guide.}

\item{LoggingConfiguration}{The Apache Airflow logs you want to send to Amazon CloudWatch Logs.}

\item{MaxWorkers}{The maximum number of workers that you want to run in your environment.
MWAA scales the number of Apache Airflow workers and the Fargate
containers that run your tasks up to the number you specify in this
field. When there are no more tasks running, and no more in the queue,
MWAA disposes of the extra containers leaving the one worker that is
included with your environment.}

\item{Name}{[required] The name of your MWAA environment.}

\item{NetworkConfiguration}{[required] The VPC networking components you want to use for your environment. At
least two private subnet identifiers and one VPC security group
identifier are required to create an environment. For more information,
see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/}{Creating the VPC network for a MWAA environment}.}

\item{PluginsS3ObjectVersion}{The \code{plugins.zip} file version you want to use.}

\item{PluginsS3Path}{The relative path to the \code{plugins.zip} file on your Amazon S3 storage
bucket. For example, \code{plugins.zip}. If a relative path is provided in
the request, then \code{PluginsS3ObjectVersion} is required. For more
information, see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/}{Importing DAGs on Amazon MWAA}.}

\item{RequirementsS3ObjectVersion}{The \code{requirements.txt} file version you want to use.}

\item{RequirementsS3Path}{The relative path to the \code{requirements.txt} file on your Amazon S3
storage bucket. For example, \code{requirements.txt}. If a relative path is
provided in the request, then \code{RequirementsS3ObjectVersion} is required.
For more information, see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/}{Importing DAGs on Amazon MWAA}.}

\item{SourceBucketArn}{[required] The Amazon Resource Name (ARN) of your Amazon S3 storage bucket. For
example, \code{arn:aws:s3:::airflow-mybucketname}.}

\item{Tags}{The metadata tags you want to attach to your environment. For more
information, see \href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS resources}.}

\item{WebserverAccessMode}{The networking access of your Apache Airflow web server. A public
network allows your Airflow UI to be accessed over the Internet by users
granted access in your IAM policy. A private network limits access of
your Airflow UI to users within your VPC. For more information, see
\href{https://docs.aws.amazon.com/mwaa/latest/userguide/}{Creating the VPC network for a MWAA environment}.}

\item{WeeklyMaintenanceWindowStart}{The day and time you want MWAA to start weekly maintenance updates on
your environment.}
}
\description{
JSON blob that describes the environment to create.
}
\section{Request syntax}{
\preformatted{svc$create_environment(
  AirflowConfigurationOptions = list(
    "string"
  ),
  AirflowVersion = "string",
  DagS3Path = "string",
  EnvironmentClass = "string",
  ExecutionRoleArn = "string",
  KmsKey = "string",
  LoggingConfiguration = list(
    DagProcessingLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    ),
    SchedulerLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    ),
    TaskLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    ),
    WebserverLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    ),
    WorkerLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    )
  ),
  MaxWorkers = 123,
  Name = "string",
  NetworkConfiguration = list(
    SecurityGroupIds = list(
      "string"
    ),
    SubnetIds = list(
      "string"
    )
  ),
  PluginsS3ObjectVersion = "string",
  PluginsS3Path = "string",
  RequirementsS3ObjectVersion = "string",
  RequirementsS3Path = "string",
  SourceBucketArn = "string",
  Tags = list(
    "string"
  ),
  WebserverAccessMode = "PRIVATE_ONLY"|"PUBLIC_ONLY",
  WeeklyMaintenanceWindowStart = "string"
)
}
}

\keyword{internal}
