% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_create_web_login_token}
\alias{mwaa_create_web_login_token}
\title{Create a JWT token to be used to login to Airflow Web UI with claims
based Authentication}
\usage{
mwaa_create_web_login_token(Name)
}
\arguments{
\item{Name}{[required] Create an Airflow Web UI login token request for a MWAA environment.}
}
\description{
Create a JWT token to be used to login to Airflow Web UI with claims
based Authentication.
}
\section{Request syntax}{
\preformatted{svc$create_web_login_token(
  Name = "string"
)
}
}

\keyword{internal}
