% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_delete_environment}
\alias{mwaa_delete_environment}
\title{Delete an existing environment}
\usage{
mwaa_delete_environment(Name)
}
\arguments{
\item{Name}{[required] The name of the environment to delete.}
}
\description{
Delete an existing environment.
}
\section{Request syntax}{
\preformatted{svc$delete_environment(
  Name = "string"
)
}
}

\keyword{internal}
