% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_get_environment}
\alias{mwaa_get_environment}
\title{Get details of an existing environment}
\usage{
mwaa_get_environment(Name)
}
\arguments{
\item{Name}{[required] The name of the environment to retrieve.}
}
\description{
Get details of an existing environment.
}
\section{Request syntax}{
\preformatted{svc$get_environment(
  Name = "string"
)
}
}

\keyword{internal}
