% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_list_environments}
\alias{mwaa_list_environments}
\title{List Amazon MWAA Environments}
\usage{
mwaa_list_environments(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum results when listing MWAA environments.}

\item{NextToken}{The Next Token when listing MWAA environments.}
}
\description{
List Amazon MWAA Environments.
}
\section{Request syntax}{
\preformatted{svc$list_environments(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
