% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_list_tags_for_resource}
\alias{mwaa_list_tags_for_resource}
\title{List the tags for MWAA environments}
\usage{
mwaa_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the MWAA environment.}
}
\description{
List the tags for MWAA environments.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
