% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_publish_metrics}
\alias{mwaa_publish_metrics}
\title{An operation for publishing metrics from the customers to the Ops plane}
\usage{
mwaa_publish_metrics(EnvironmentName, MetricData)
}
\arguments{
\item{EnvironmentName}{[required] Publishes environment metric data to Amazon CloudWatch.}

\item{MetricData}{[required] Publishes metric data points to Amazon CloudWatch. CloudWatch associates
the data points with the specified metrica.}
}
\description{
An operation for publishing metrics from the customers to the Ops plane.
}
\section{Request syntax}{
\preformatted{svc$publish_metrics(
  EnvironmentName = "string",
  MetricData = list(
    list(
      Dimensions = list(
        list(
          Name = "string",
          Value = "string"
        )
      ),
      MetricName = "string",
      StatisticValues = list(
        Maximum = 123.0,
        Minimum = 123.0,
        SampleCount = 123,
        Sum = 123.0
      ),
      Timestamp = as.POSIXct(
        "2015-01-01"
      ),
      Unit = "Seconds"|"Microseconds"|"Milliseconds"|"Bytes"|"Kilobytes"|"Megabytes"|"Gigabytes"|"Terabytes"|"Bits"|"Kilobits"|"Megabits"|"Gigabits"|"Terabits"|"Percent"|"Count"|"Bytes/Second"|"Kilobytes/Second"|"Megabytes/Second"|"Gigabytes/Second"|"Terabytes/Second"|"Bits/Second"|"Kilobits/Second"|"Megabits/Second"|"Gigabits/Second"|"Terabits/Second"|"Count/Second"|"None",
      Value = 123.0
    )
  )
)
}
}

\keyword{internal}
