% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_tag_resource}
\alias{mwaa_tag_resource}
\title{Add tag to the MWAA environments}
\usage{
mwaa_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The tag resource ARN of the MWAA environments.}

\item{Tags}{[required] The tag resource tag of the MWAA environments.}
}
\description{
Add tag to the MWAA environments.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
