% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_untag_resource}
\alias{mwaa_untag_resource}
\title{Remove a tag from the MWAA environments}
\usage{
mwaa_untag_resource(ResourceArn, tagKeys)
}
\arguments{
\item{ResourceArn}{[required] The tag resource ARN of the MWAA environments.}

\item{tagKeys}{[required] The tag resource key of the MWAA environments.}
}
\description{
Remove a tag from the MWAA environments.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
