% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_update_environment}
\alias{mwaa_update_environment}
\title{Update an MWAA environment}
\usage{
mwaa_update_environment(AirflowConfigurationOptions, AirflowVersion,
  DagS3Path, EnvironmentClass, ExecutionRoleArn, LoggingConfiguration,
  MaxWorkers, Name, NetworkConfiguration, PluginsS3ObjectVersion,
  PluginsS3Path, RequirementsS3ObjectVersion, RequirementsS3Path,
  SourceBucketArn, WebserverAccessMode, WeeklyMaintenanceWindowStart)
}
\arguments{
\item{AirflowConfigurationOptions}{The Airflow Configuration Options to update of your Amazon MWAA
environment.}

\item{AirflowVersion}{The Airflow Version to update of your Amazon MWAA environment.}

\item{DagS3Path}{The Dags folder S3 Path to update of your Amazon MWAA environment.}

\item{EnvironmentClass}{The Environment Class to update of your Amazon MWAA environment.}

\item{ExecutionRoleArn}{The Executio Role ARN to update of your Amazon MWAA environment.}

\item{LoggingConfiguration}{The Logging Configuration to update of your Amazon MWAA environment.}

\item{MaxWorkers}{The Maximum Workers to update of your Amazon MWAA environment.}

\item{Name}{[required] The name of your Amazon MWAA environment that you wish to update.}

\item{NetworkConfiguration}{The Network Configuration to update of your Amazon MWAA environment.}

\item{PluginsS3ObjectVersion}{The Plugins.zip S3 Object Version to update of your Amazon MWAA
environment.}

\item{PluginsS3Path}{The Plugins.zip S3 Path to update of your Amazon MWAA environment.}

\item{RequirementsS3ObjectVersion}{The Requirements.txt S3 ObjectV ersion to update of your Amazon MWAA
environment.}

\item{RequirementsS3Path}{The Requirements.txt S3 Path to update of your Amazon MWAA environment.}

\item{SourceBucketArn}{The S3 Source Bucket ARN to update of your Amazon MWAA environment.}

\item{WebserverAccessMode}{The Webserver Access Mode to update of your Amazon MWAA environment.}

\item{WeeklyMaintenanceWindowStart}{The Weekly Maintenance Window Start to update of your Amazon MWAA
environment.}
}
\description{
Update an MWAA environment.
}
\section{Request syntax}{
\preformatted{svc$update_environment(
  AirflowConfigurationOptions = list(
    "string"
  ),
  AirflowVersion = "string",
  DagS3Path = "string",
  EnvironmentClass = "string",
  ExecutionRoleArn = "string",
  LoggingConfiguration = list(
    DagProcessingLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    ),
    SchedulerLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    ),
    TaskLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    ),
    WebserverLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    ),
    WorkerLogs = list(
      Enabled = TRUE|FALSE,
      LogLevel = "CRITICAL"|"ERROR"|"WARNING"|"INFO"|"DEBUG"
    )
  ),
  MaxWorkers = 123,
  Name = "string",
  NetworkConfiguration = list(
    SecurityGroupIds = list(
      "string"
    )
  ),
  PluginsS3ObjectVersion = "string",
  PluginsS3Path = "string",
  RequirementsS3ObjectVersion = "string",
  RequirementsS3Path = "string",
  SourceBucketArn = "string",
  WebserverAccessMode = "PRIVATE_ONLY"|"PUBLIC_ONLY",
  WeeklyMaintenanceWindowStart = "string"
)
}
}

\keyword{internal}
