% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_delete_db_instance}
\alias{neptune_delete_db_instance}
\title{The DeleteDBInstance action deletes a previously provisioned DB instance}
\usage{
neptune_delete_db_instance(DBInstanceIdentifier, SkipFinalSnapshot,
  FinalDBSnapshotIdentifier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The DB instance identifier for the DB instance to be deleted. This
parameter isn't case-sensitive.

Constraints:
\itemize{
\item Must match the name of an existing DB instance.
}}

\item{SkipFinalSnapshot}{Determines whether a final DB snapshot is created before the DB instance
is deleted. If \code{true} is specified, no DBSnapshot is created. If \code{false}
is specified, a DB snapshot is created before the DB instance is
deleted.

Note that when a DB instance is in a failure state and has a status of
'failed', 'incompatible-restore', or 'incompatible-network', it
can only be deleted when the SkipFinalSnapshot parameter is set to
"true".

Specify \code{true} when deleting a Read Replica.

The FinalDBSnapshotIdentifier parameter must be specified if
SkipFinalSnapshot is \code{false}.

Default: \code{false}}

\item{FinalDBSnapshotIdentifier}{The DBSnapshotIdentifier of the new DBSnapshot created when
SkipFinalSnapshot is set to \code{false}.

Specifying this parameter and also setting the SkipFinalShapshot
parameter to true results in an error.

Constraints:
\itemize{
\item Must be 1 to 255 letters or numbers.
\item First character must be a letter
\item Cannot end with a hyphen or contain two consecutive hyphens
\item Cannot be specified when deleting a Read Replica.
}}
}
\description{
The DeleteDBInstance action deletes a previously provisioned DB
instance. When you delete a DB instance, all automated backups for that
instance are deleted and can't be recovered. Manual DB snapshots of the
DB instance to be deleted by \code{DeleteDBInstance} are not deleted.
}
\details{
If you request a final DB snapshot the status of the Amazon Neptune DB
instance is \code{deleting} until the DB snapshot is created. The API action
\code{DescribeDBInstance} is used to monitor the status of this operation.
The action can't be canceled or reverted once submitted.

Note that when a DB instance is in a failure state and has a status of
\code{failed}, \code{incompatible-restore}, or \code{incompatible-network}, you can
only delete it when the \code{SkipFinalSnapshot} parameter is set to \code{true}.

If the specified DB instance is part of a DB cluster, you can't delete
the DB instance if both of the following conditions are true:
\itemize{
\item The DB instance is the only instance in the DB cluster.
}
}
\section{Request syntax}{
\preformatted{svc$delete_db_instance(
  DBInstanceIdentifier = "string",
  SkipFinalSnapshot = TRUE|FALSE,
  FinalDBSnapshotIdentifier = "string"
)
}
}

\keyword{internal}
