% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_delete_event_subscription}
\alias{neptune_delete_event_subscription}
\title{Deletes an event notification subscription}
\usage{
neptune_delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{[required] The name of the event notification subscription you want to delete.}
}
\description{
Deletes an event notification subscription.
}
\section{Request syntax}{
\preformatted{neptune$delete_event_subscription(
  SubscriptionName = "string"
)
}
}

\keyword{internal}
