% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_promote_read_replica_db_cluster}
\alias{neptune_promote_read_replica_db_cluster}
\title{Not supported}
\usage{
neptune_promote_read_replica_db_cluster(DBClusterIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{[required] Not supported.}
}
\description{
Not supported.
}
\section{Request syntax}{
\preformatted{svc$promote_read_replica_db_cluster(
  DBClusterIdentifier = "string"
)
}
}

\keyword{internal}
