% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_stop_db_cluster}
\alias{neptune_stop_db_cluster}
\title{Stops an Amazon Neptune DB cluster}
\usage{
neptune_stop_db_cluster(DBClusterIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{[required] The DB cluster identifier of the Neptune DB cluster to be stopped. This
parameter is stored as a lowercase string.}
}
\description{
Stops an Amazon Neptune DB cluster. When you stop a DB cluster, Neptune
retains the DB cluster's metadata, including its endpoints and DB
parameter groups.
}
\details{
Neptune also retains the transaction logs so you can do a point-in-time
restore if necessary.
}
\section{Request syntax}{
\preformatted{svc$stop_db_cluster(
  DBClusterIdentifier = "string"
)
}
}

\keyword{internal}
