% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{networkfirewall}
\alias{networkfirewall}
\title{AWS Network Firewall}
\usage{
networkfirewall(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
\item{\strong{use_dual_stack}: Set this to \code{true} to use the dualstack (IPv4 and IPv6) endpoint for a service, where available, falling back to the regular endpoint when it isn't. Defaults to the \code{AWS_USE_DUALSTACK_ENDPOINT} environment variable when unset.}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
This is the API Reference for Network Firewall. This guide is for developers who need detailed information about the Network Firewall API actions, data types, and errors.

The REST API requires you to handle connection details, such as calculating signatures, handling request retries, and error handling. For general information about using the Amazon Web Services REST APIs, see \href{https://docs.aws.amazon.com/general/latest/gr/}{Amazon Web Services APIs}.

To view the complete list of Amazon Web Services Regions where Network Firewall is available, see \href{https://docs.aws.amazon.com/general/latest/gr/network-firewall.html}{Service endpoints and quotas} in the \emph{Amazon Web Services General Reference}.

To access Network Firewall using the IPv4 REST API endpoint: \verb{https://network-firewall.<region>.amazonaws.com }

To access Network Firewall using the Dualstack (IPv4 and IPv6) REST API endpoint: \verb{https://network-firewall.<region>.aws.api }

Alternatively, you can use one of the Amazon Web Services SDKs to access an API that's tailored to the programming language or platform that you're using. For more information, see \href{https://builder.aws.com/build/tools#SDKs}{Amazon Web Services SDKs}.

For descriptions of Network Firewall features, including and step-by-step instructions on how to use them through the Network Firewall console, see the \href{https://docs.aws.amazon.com/network-firewall/latest/developerguide/}{Network Firewall Developer Guide}.

Network Firewall is a stateful, managed, network firewall and intrusion detection and prevention service for Amazon Virtual Private Cloud (Amazon VPC). With Network Firewall, you can filter traffic at the perimeter of your VPC. This includes filtering traffic going to and coming from an internet gateway, NAT gateway, or over VPN or Direct Connect. Network Firewall uses rules that are compatible with Suricata, a free, open source network analysis and threat detection engine. Network Firewall supports Suricata version 7.0.8. For information about Suricata, see the \href{https://suricata.io/}{Suricata website} and the \href{https://docs.suricata.io/en/suricata-7.0.8/}{Suricata User Guide}.

You can use Network Firewall to monitor and protect your VPC traffic in a number of ways. The following are just a few examples:
\itemize{
\item Allow domains or IP addresses for known Amazon Web Services service endpoints, such as Amazon S3, and block all other forms of traffic.
\item Use custom lists of known bad domains to limit the types of domain names that your applications can access.
\item Perform deep packet inspection on traffic entering or leaving your VPC.
\item Use stateful protocol detection to filter protocols like HTTPS, regardless of the port used.
}

To enable Network Firewall for your VPCs, you perform steps in both Amazon VPC and in Network Firewall. For information about using Amazon VPC, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/}{Amazon VPC User Guide}.

To start using Network Firewall, do the following:
\enumerate{
\item (Optional) If you don't already have a VPC that you want to protect, create it in Amazon VPC.
\item In Amazon VPC, in each Availability Zone where you want to have a firewall endpoint, create a subnet for the sole use of Network Firewall.
\item In Network Firewall, define the firewall behavior as follows:
\enumerate{
\item Create stateless and stateful rule groups, to define the components of the network traffic filtering behavior that you want your firewall to have.
\item Create a firewall policy that uses your rule groups and specifies additional default traffic filtering behavior.
}
\item In Network Firewall, create a firewall and specify your new firewall policy and VPC subnets. Network Firewall creates a firewall endpoint in each subnet that you specify, with the behavior that's defined in the firewall policy.
\item In Amazon VPC, use ingress routing enhancements to route traffic through the new firewall endpoints.
}

After your firewall is established, you can add firewall endpoints for new Availability Zones by following the prior steps for the Amazon VPC setup and firewall subnet definitions. You can also add endpoints to Availability Zones that you're using in the firewall, either for the same VPC or for another VPC, by following the prior steps for the Amazon VPC setup, and defining the new VPC subnets as VPC endpoint associations.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- networkfirewall(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string",
    use_dual_stack = "logical"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.networking:networkfirewall_accept_network_firewall_transit_gateway_attachment]{accept_network_firewall_transit_gateway_attachment} \tab Accepts a transit gateway attachment request for Network Firewall\cr
\link[paws.networking:networkfirewall_associate_availability_zones]{associate_availability_zones} \tab Associates the specified Availability Zones with a transit gateway-attached firewall\cr
\link[paws.networking:networkfirewall_associate_firewall_policy]{associate_firewall_policy} \tab Associates a FirewallPolicy to a Firewall\cr
\link[paws.networking:networkfirewall_associate_subnets]{associate_subnets} \tab Associates the specified subnets in the Amazon VPC to the firewall\cr
\link[paws.networking:networkfirewall_attach_rule_groups_to_proxy_configuration]{attach_rule_groups_to_proxy_configuration} \tab Attaches ProxyRuleGroup resources to a ProxyConfiguration\cr
\link[paws.networking:networkfirewall_create_container_association]{create_container_association} \tab Creates a Network Firewall container association\cr
\link[paws.networking:networkfirewall_create_firewall]{create_firewall} \tab Creates an Network Firewall Firewall and accompanying FirewallStatus for a VPC\cr
\link[paws.networking:networkfirewall_create_firewall_policy]{create_firewall_policy} \tab Creates the firewall policy for the firewall according to the specifications\cr
\link[paws.networking:networkfirewall_create_proxy]{create_proxy} \tab Creates an Network Firewall Proxy\cr
\link[paws.networking:networkfirewall_create_proxy_configuration]{create_proxy_configuration} \tab Creates an Network Firewall ProxyConfiguration\cr
\link[paws.networking:networkfirewall_create_proxy_rule_group]{create_proxy_rule_group} \tab Creates an Network Firewall ProxyRuleGroup\cr
\link[paws.networking:networkfirewall_create_proxy_rules]{create_proxy_rules} \tab Creates Network Firewall ProxyRule resources\cr
\link[paws.networking:networkfirewall_create_rule_group]{create_rule_group} \tab Creates the specified stateless or stateful rule group, which includes the rules for network traffic inspection, a capacity setting, and tags\cr
\link[paws.networking:networkfirewall_create_tls_inspection_configuration]{create_tls_inspection_configuration} \tab Creates an Network Firewall TLS inspection configuration\cr
\link[paws.networking:networkfirewall_create_vpc_endpoint_association]{create_vpc_endpoint_association} \tab Creates a firewall endpoint for an Network Firewall firewall\cr
\link[paws.networking:networkfirewall_delete_container_association]{delete_container_association} \tab Deletes a container association\cr
\link[paws.networking:networkfirewall_delete_firewall]{delete_firewall} \tab Deletes the specified Firewall and its FirewallStatus\cr
\link[paws.networking:networkfirewall_delete_firewall_policy]{delete_firewall_policy} \tab Deletes the specified FirewallPolicy\cr
\link[paws.networking:networkfirewall_delete_network_firewall_transit_gateway_attachment]{delete_network_firewall_transit_gateway_attachment} \tab Deletes a transit gateway attachment from a Network Firewall\cr
\link[paws.networking:networkfirewall_delete_proxy]{delete_proxy} \tab Deletes the specified Proxy\cr
\link[paws.networking:networkfirewall_delete_proxy_configuration]{delete_proxy_configuration} \tab Deletes the specified ProxyConfiguration\cr
\link[paws.networking:networkfirewall_delete_proxy_rule_group]{delete_proxy_rule_group} \tab Deletes the specified ProxyRuleGroup\cr
\link[paws.networking:networkfirewall_delete_proxy_rules]{delete_proxy_rules} \tab Deletes the specified ProxyRule(s)\cr
\link[paws.networking:networkfirewall_delete_resource_policy]{delete_resource_policy} \tab Deletes a resource policy that you created in a PutResourcePolicy request\cr
\link[paws.networking:networkfirewall_delete_rule_group]{delete_rule_group} \tab Deletes the specified RuleGroup\cr
\link[paws.networking:networkfirewall_delete_tls_inspection_configuration]{delete_tls_inspection_configuration} \tab Deletes the specified TLSInspectionConfiguration\cr
\link[paws.networking:networkfirewall_delete_vpc_endpoint_association]{delete_vpc_endpoint_association} \tab Deletes the specified VpcEndpointAssociation\cr
\link[paws.networking:networkfirewall_describe_container_association]{describe_container_association} \tab Retrieves the configuration and status of a container association\cr
\link[paws.networking:networkfirewall_describe_firewall]{describe_firewall} \tab Returns the data objects for the specified firewall\cr
\link[paws.networking:networkfirewall_describe_firewall_metadata]{describe_firewall_metadata} \tab Returns the high-level information about a firewall, including the Availability Zones where the Firewall is currently in use\cr
\link[paws.networking:networkfirewall_describe_firewall_policy]{describe_firewall_policy} \tab Returns the data objects for the specified firewall policy\cr
\link[paws.networking:networkfirewall_describe_flow_operation]{describe_flow_operation} \tab Returns key information about a specific flow operation\cr
\link[paws.networking:networkfirewall_describe_logging_configuration]{describe_logging_configuration} \tab Returns the logging configuration for the specified firewall\cr
\link[paws.networking:networkfirewall_describe_proxy]{describe_proxy} \tab Returns the data objects for the specified proxy\cr
\link[paws.networking:networkfirewall_describe_proxy_configuration]{describe_proxy_configuration} \tab Returns the data objects for the specified proxy configuration\cr
\link[paws.networking:networkfirewall_describe_proxy_rule]{describe_proxy_rule} \tab Returns the data objects for the specified proxy configuration for the specified proxy rule group\cr
\link[paws.networking:networkfirewall_describe_proxy_rule_group]{describe_proxy_rule_group} \tab Returns the data objects for the specified proxy rule group\cr
\link[paws.networking:networkfirewall_describe_resource_policy]{describe_resource_policy} \tab Retrieves a resource policy that you created in a PutResourcePolicy request\cr
\link[paws.networking:networkfirewall_describe_rule_group]{describe_rule_group} \tab Returns the data objects for the specified rule group\cr
\link[paws.networking:networkfirewall_describe_rule_group_metadata]{describe_rule_group_metadata} \tab High-level information about a rule group, returned by operations like create and describe\cr
\link[paws.networking:networkfirewall_describe_rule_group_summary]{describe_rule_group_summary} \tab Returns detailed information for a stateful rule group\cr
\link[paws.networking:networkfirewall_describe_tls_inspection_configuration]{describe_tls_inspection_configuration} \tab Returns the data objects for the specified TLS inspection configuration\cr
\link[paws.networking:networkfirewall_describe_vpc_endpoint_association]{describe_vpc_endpoint_association} \tab Returns the data object for the specified VPC endpoint association\cr
\link[paws.networking:networkfirewall_detach_rule_groups_from_proxy_configuration]{detach_rule_groups_from_proxy_configuration} \tab Detaches ProxyRuleGroup resources from a ProxyConfiguration\cr
\link[paws.networking:networkfirewall_disassociate_availability_zones]{disassociate_availability_zones} \tab Removes the specified Availability Zone associations from a transit gateway-attached firewall\cr
\link[paws.networking:networkfirewall_disassociate_subnets]{disassociate_subnets} \tab Removes the specified subnet associations from the firewall\cr
\link[paws.networking:networkfirewall_get_analysis_report_results]{get_analysis_report_results} \tab The results of a COMPLETED analysis report generated with StartAnalysisReport\cr
\link[paws.networking:networkfirewall_list_analysis_reports]{list_analysis_reports} \tab Returns a list of all traffic analysis reports generated within the last 30 days\cr
\link[paws.networking:networkfirewall_list_container_associations]{list_container_associations} \tab Lists the container associations in your account and Region\cr
\link[paws.networking:networkfirewall_list_firewall_policies]{list_firewall_policies} \tab Retrieves the metadata for the firewall policies that you have defined\cr
\link[paws.networking:networkfirewall_list_firewalls]{list_firewalls} \tab Retrieves the metadata for the firewalls that you have defined\cr
\link[paws.networking:networkfirewall_list_flow_operation_results]{list_flow_operation_results} \tab Returns the results of a specific flow operation\cr
\link[paws.networking:networkfirewall_list_flow_operations]{list_flow_operations} \tab Returns a list of all flow operations ran in a specific firewall\cr
\link[paws.networking:networkfirewall_list_proxies]{list_proxies} \tab Retrieves the metadata for the proxies that you have defined\cr
\link[paws.networking:networkfirewall_list_proxy_configurations]{list_proxy_configurations} \tab Retrieves the metadata for the proxy configuration that you have defined\cr
\link[paws.networking:networkfirewall_list_proxy_rule_groups]{list_proxy_rule_groups} \tab Retrieves the metadata for the proxy rule groups that you have defined\cr
\link[paws.networking:networkfirewall_list_rule_groups]{list_rule_groups} \tab Retrieves the metadata for the rule groups that you have defined\cr
\link[paws.networking:networkfirewall_list_tags_for_resource]{list_tags_for_resource} \tab Retrieves the tags associated with the specified resource\cr
\link[paws.networking:networkfirewall_list_tls_inspection_configurations]{list_tls_inspection_configurations} \tab Retrieves the metadata for the TLS inspection configurations that you have defined\cr
\link[paws.networking:networkfirewall_list_vpc_endpoint_associations]{list_vpc_endpoint_associations} \tab Retrieves the metadata for the VPC endpoint associations that you have defined\cr
\link[paws.networking:networkfirewall_put_resource_policy]{put_resource_policy} \tab Creates or updates an IAM policy for your rule group, firewall policy, or firewall\cr
\link[paws.networking:networkfirewall_reject_network_firewall_transit_gateway_attachment]{reject_network_firewall_transit_gateway_attachment} \tab Rejects a transit gateway attachment request for Network Firewall\cr
\link[paws.networking:networkfirewall_start_analysis_report]{start_analysis_report} \tab Generates a traffic analysis report for the timeframe and traffic type you specify\cr
\link[paws.networking:networkfirewall_start_flow_capture]{start_flow_capture} \tab Begins capturing the flows in a firewall, according to the filters you define\cr
\link[paws.networking:networkfirewall_start_flow_flush]{start_flow_flush} \tab Begins the flushing of traffic from the firewall, according to the filters you define\cr
\link[paws.networking:networkfirewall_tag_resource]{tag_resource} \tab Adds the specified tags to the specified resource\cr
\link[paws.networking:networkfirewall_untag_resource]{untag_resource} \tab Removes the tags with the specified keys from the specified resource\cr
\link[paws.networking:networkfirewall_update_availability_zone_change_protection]{update_availability_zone_change_protection} \tab Modifies the AvailabilityZoneChangeProtection setting for a transit gateway-attached firewall\cr
\link[paws.networking:networkfirewall_update_container_association]{update_container_association} \tab Updates the monitoring configurations and description of a container association\cr
\link[paws.networking:networkfirewall_update_firewall_analysis_settings]{update_firewall_analysis_settings} \tab Enables specific types of firewall analysis on a specific firewall you define\cr
\link[paws.networking:networkfirewall_update_firewall_delete_protection]{update_firewall_delete_protection} \tab Modifies the flag, DeleteProtection, which indicates whether it is possible to delete the firewall\cr
\link[paws.networking:networkfirewall_update_firewall_description]{update_firewall_description} \tab Modifies the description for the specified firewall\cr
\link[paws.networking:networkfirewall_update_firewall_encryption_configuration]{update_firewall_encryption_configuration} \tab A complex type that contains settings for encryption of your firewall resources\cr
\link[paws.networking:networkfirewall_update_firewall_policy]{update_firewall_policy} \tab Updates the properties of the specified firewall policy\cr
\link[paws.networking:networkfirewall_update_firewall_policy_change_protection]{update_firewall_policy_change_protection} \tab Modifies the flag, ChangeProtection, which indicates whether it is possible to change the firewall\cr
\link[paws.networking:networkfirewall_update_logging_configuration]{update_logging_configuration} \tab Sets the logging configuration for the specified firewall\cr
\link[paws.networking:networkfirewall_update_proxy]{update_proxy} \tab Updates the properties of the specified proxy\cr
\link[paws.networking:networkfirewall_update_proxy_configuration]{update_proxy_configuration} \tab Updates the properties of the specified proxy configuration\cr
\link[paws.networking:networkfirewall_update_proxy_rule]{update_proxy_rule} \tab Updates the properties of the specified proxy rule\cr
\link[paws.networking:networkfirewall_update_proxy_rule_group_priorities]{update_proxy_rule_group_priorities} \tab Updates proxy rule group priorities within a proxy configuration\cr
\link[paws.networking:networkfirewall_update_proxy_rule_priorities]{update_proxy_rule_priorities} \tab Updates proxy rule priorities within a proxy rule group\cr
\link[paws.networking:networkfirewall_update_proxy_settings]{update_proxy_settings} \tab Modifies the proxy listener configuration of a proxy mode firewall\cr
\link[paws.networking:networkfirewall_update_rule_group]{update_rule_group} \tab Updates the rule settings for the specified rule group\cr
\link[paws.networking:networkfirewall_update_subnet_change_protection]{update_subnet_change_protection} \tab Update subnet change protection\cr
\link[paws.networking:networkfirewall_update_tls_inspection_configuration]{update_tls_inspection_configuration} \tab Updates the TLS inspection configuration settings for the specified TLS inspection configuration
}
}

\examples{
\dontrun{
svc <- networkfirewall()
svc$accept_network_firewall_transit_gateway_attachment(
  Foo = 123
)
}

}
