% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_associate_firewall_policy}
\alias{networkfirewall_associate_firewall_policy}
\title{Associates a FirewallPolicy to a Firewall}
\usage{
networkfirewall_associate_firewall_policy(UpdateToken, FirewallArn,
  FirewallName, FirewallPolicyArn)
}
\arguments{
\item{UpdateToken}{An optional token that you can use for optimistic locking. Network
Firewall returns a token to your requests that access the firewall. The
token marks the state of the firewall resource at the time of the
request.

To make an unconditional change to the firewall, omit the token in your
update request. Without the token, Network Firewall performs your
updates regardless of whether the firewall has changed since you last
retrieved it.

To make a conditional change to the firewall, provide the token in your
update request. Network Firewall uses the token to ensure that the
firewall hasn't changed since you last retrieved it. If it has changed,
the operation fails with an \code{InvalidTokenException}. If this happens,
retrieve the firewall again to get a current copy of it with a new
token. Reapply your changes as needed, then try the operation again
using the new token.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallPolicyArn}{[required] The Amazon Resource Name (ARN) of the firewall policy.}
}
\description{
Associates a FirewallPolicy to a Firewall.

A firewall policy defines how to monitor and manage your VPC network
traffic, using a collection of inspection rule groups and other
settings. Each firewall requires one firewall policy association, and
you can use the same firewall policy for multiple firewalls.
}
\section{Request syntax}{
\preformatted{svc$associate_firewall_policy(
  UpdateToken = "string",
  FirewallArn = "string",
  FirewallName = "string",
  FirewallPolicyArn = "string"
)
}
}

\keyword{internal}
