% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_create_firewall}
\alias{networkfirewall_create_firewall}
\title{Creates an AWS Network Firewall Firewall and accompanying FirewallStatus
for a VPC}
\usage{
networkfirewall_create_firewall(FirewallName, FirewallPolicyArn, VpcId,
  SubnetMappings, DeleteProtection, SubnetChangeProtection,
  FirewallPolicyChangeProtection, Description, Tags)
}
\arguments{
\item{FirewallName}{[required] The descriptive name of the firewall. You can't change the name of a
firewall after you create it.}

\item{FirewallPolicyArn}{[required] The Amazon Resource Name (ARN) of the FirewallPolicy that you want to
use for the firewall.}

\item{VpcId}{[required] The unique identifier of the VPC where Network Firewall should create
the firewall.

You can't change this setting after you create the firewall.}

\item{SubnetMappings}{[required] The public subnets to use for your Network Firewall firewalls. Each
subnet must belong to a different Availability Zone in the VPC. Network
Firewall creates a firewall endpoint in each subnet.}

\item{DeleteProtection}{A flag indicating whether it is possible to delete the firewall. A
setting of \code{TRUE} indicates that the firewall is protected against
deletion. Use this setting to protect against accidentally deleting a
firewall that is in use. When you create a firewall, the operation
initializes this flag to \code{TRUE}.}

\item{SubnetChangeProtection}{A setting indicating whether the firewall is protected against changes
to the subnet associations. Use this setting to protect against
accidentally modifying the subnet associations for a firewall that is in
use. When you create a firewall, the operation initializes this setting
to \code{TRUE}.}

\item{FirewallPolicyChangeProtection}{A setting indicating whether the firewall is protected against a change
to the firewall policy association. Use this setting to protect against
accidentally modifying the firewall policy for a firewall that is in
use. When you create a firewall, the operation initializes this setting
to \code{TRUE}.}

\item{Description}{A description of the firewall.}

\item{Tags}{The key:value pairs to associate with the resource.}
}
\description{
Creates an AWS Network Firewall Firewall and accompanying FirewallStatus
for a VPC.

The firewall defines the configuration settings for an AWS Network
Firewall firewall. The settings that you can define at creation include
the firewall policy, the subnets in your VPC to use for the firewall
endpoints, and any tags that are attached to the firewall AWS resource.

After you create a firewall, you can provide additional settings, like
the logging configuration.

To update the settings for a firewall, you use the operations that apply
to the settings themselves, for example
\code{\link[=networkfirewall_update_logging_configuration]{update_logging_configuration}},
\code{\link[=networkfirewall_associate_subnets]{associate_subnets}}, and
\code{\link[=networkfirewall_update_firewall_delete_protection]{update_firewall_delete_protection}}.

To manage a firewall's tags, use the standard AWS resource tagging
operations,
\code{\link[=networkfirewall_list_tags_for_resource]{list_tags_for_resource}},
\code{\link[=networkfirewall_tag_resource]{tag_resource}}, and
\code{\link[=networkfirewall_untag_resource]{untag_resource}}.

To retrieve information about firewalls, use
\code{\link[=networkfirewall_list_firewalls]{list_firewalls}} and
\code{\link[=networkfirewall_describe_firewall]{describe_firewall}}.
}
\section{Request syntax}{
\preformatted{svc$create_firewall(
  FirewallName = "string",
  FirewallPolicyArn = "string",
  VpcId = "string",
  SubnetMappings = list(
    list(
      SubnetId = "string"
    )
  ),
  DeleteProtection = TRUE|FALSE,
  SubnetChangeProtection = TRUE|FALSE,
  FirewallPolicyChangeProtection = TRUE|FALSE,
  Description = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
