% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_firewall}
\alias{networkfirewall_delete_firewall}
\title{Deletes the specified Firewall and its FirewallStatus}
\usage{
networkfirewall_delete_firewall(FirewallName, FirewallArn)
}
\arguments{
\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Deletes the specified Firewall and its FirewallStatus. This operation
requires the firewall's \code{DeleteProtection} flag to be \code{FALSE}. You can't
revert this operation.

You can check whether a firewall is in use by reviewing the route tables
for the Availability Zones where you have firewall subnet mappings.
Retrieve the subnet mappings by calling DescribeFirewall. You define and
update the route tables through Amazon VPC. As needed, update the route
tables for the zones to remove the firewall endpoints. When the route
tables no longer use the firewall endpoints, you can remove the firewall
safely.

To delete a firewall, remove the delete protection if you need to using
UpdateFirewallDeleteProtection, then delete the firewall by calling
DeleteFirewall.
}
\section{Request syntax}{
\preformatted{svc$delete_firewall(
  FirewallName = "string",
  FirewallArn = "string"
)
}
}

\keyword{internal}
