% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_resource_policy}
\alias{networkfirewall_delete_resource_policy}
\title{Deletes a resource policy that you created in a PutResourcePolicy
request}
\usage{
networkfirewall_delete_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the rule group or firewall policy
whose resource policy you want to delete.}
}
\description{
Deletes a resource policy that you created in a
\code{\link[=networkfirewall_put_resource_policy]{put_resource_policy}} request.
}
\section{Request syntax}{
\preformatted{svc$delete_resource_policy(
  ResourceArn = "string"
)
}
}

\keyword{internal}
