% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_firewall}
\alias{networkfirewall_describe_firewall}
\title{Returns the data objects for the specified firewall}
\usage{
networkfirewall_describe_firewall(FirewallName, FirewallArn)
}
\arguments{
\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Returns the data objects for the specified firewall.
}
\section{Request syntax}{
\preformatted{svc$describe_firewall(
  FirewallName = "string",
  FirewallArn = "string"
)
}
}

\keyword{internal}
