% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_put_resource_policy}
\alias{networkfirewall_put_resource_policy}
\title{Creates or updates an AWS Identity and Access Management policy for your
rule group or firewall policy}
\usage{
networkfirewall_put_resource_policy(ResourceArn, Policy)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the account that you want to share
rule groups and firewall policies with.}

\item{Policy}{[required] The AWS Identity and Access Management policy statement that lists the
accounts that you want to share your rule group or firewall policy with
and the operations that you want the accounts to be able to perform.

For a rule group resource, you can specify the following operations in
the Actions section of the statement:
\itemize{
\item network-firewall:CreateFirewallPolicy
\item network-firewall:UpdateFirewallPolicy
\item network-firewall:ListRuleGroups
}

For a firewall policy resource, you can specify the following operations
in the Actions section of the statement:
\itemize{
\item network-firewall:CreateFirewall
\item network-firewall:UpdateFirewall
\item network-firewall:AssociateFirewallPolicy
\item network-firewall:ListFirewallPolicies
}

In the Resource section of the statement, you specify the ARNs for the
rule groups and firewall policies that you want to share with the
account that you specified in \code{Arn}.}
}
\description{
Creates or updates an AWS Identity and Access Management policy for your
rule group or firewall policy. Use this to share rule groups and
firewall policies between accounts. This operation works in conjunction
with the AWS Resource Access Manager (RAM) service to manage resource
sharing for Network Firewall.

Use this operation to create or update a resource policy for your rule
group or firewall policy. In the policy, you specify the accounts that
you want to share the resource with and the operations that you want the
accounts to be able to perform.

When you add an account in the resource policy, you then run the
following Resource Access Manager (RAM) operations to access and accept
the shared rule group or firewall policy.
\itemize{
\item \href{https://docs.aws.amazon.com/ram/latest/APIReference/API_GetResourceShareInvitations.html}{GetResourceShareInvitations} -
Returns the Amazon Resource Names (ARNs) of the resource share
invitations.
\item \href{https://docs.aws.amazon.com/ram/latest/APIReference/API_AcceptResourceShareInvitation.html}{AcceptResourceShareInvitation} -
Accepts the share invitation for a specified resource share.
}

For additional information about resource sharing using RAM, see \href{https://docs.aws.amazon.com/ram/latest/userguide/what-is.html}{AWS Resource Access Manager User Guide}.
}
\section{Request syntax}{
\preformatted{svc$put_resource_policy(
  ResourceArn = "string",
  Policy = "string"
)
}
}

\keyword{internal}
