% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_update_logging_configuration}
\alias{networkfirewall_update_logging_configuration}
\title{Sets the logging configuration for the specified firewall}
\usage{
networkfirewall_update_logging_configuration(FirewallArn, FirewallName,
  LoggingConfiguration)
}
\arguments{
\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{LoggingConfiguration}{Defines how Network Firewall performs logging for a firewall. If you
omit this setting, Network Firewall disables logging for the firewall.}
}
\description{
Sets the logging configuration for the specified firewall.

To change the logging configuration, retrieve the LoggingConfiguration
by calling DescribeLoggingConfiguration, then change it and provide the
modified object to this update call. You must change the logging
configuration one LogDestinationConfig at a time inside the retrieved
LoggingConfiguration object.

You can perform only one of the following actions in any call to
\code{UpdateLoggingConfiguration}:
\itemize{
\item Create a new log destination object by adding a single
\code{LogDestinationConfig} array element to \code{LogDestinationConfigs}.
\item Delete a log destination object by removing a single
\code{LogDestinationConfig} array element from \code{LogDestinationConfigs}.
\item Change the \code{LogDestination} setting in a single
\code{LogDestinationConfig} array element.
}

You can't change the \code{LogDestinationType} or \code{LogType} in a
\code{LogDestinationConfig}. To change these settings, delete the existing
\code{LogDestinationConfig} object and create a new one, using two separate
calls to this update operation.
}
\section{Request syntax}{
\preformatted{svc$update_logging_configuration(
  FirewallArn = "string",
  FirewallName = "string",
  LoggingConfiguration = list(
    LogDestinationConfigs = list(
      list(
        LogType = "ALERT"|"FLOW",
        LogDestinationType = "S3"|"CloudWatchLogs"|"KinesisDataFirehose",
        LogDestination = list(
          "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
