% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_associate_customer_gateway}
\alias{networkmanager_associate_customer_gateway}
\title{Associates a customer gateway with a device and optionally, with a link}
\usage{
networkmanager_associate_customer_gateway(CustomerGatewayArn,
  GlobalNetworkId, DeviceId, LinkId)
}
\arguments{
\item{CustomerGatewayArn}{[required] The Amazon Resource Name (ARN) of the customer gateway. For more
information, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/list_amazonec2.html#amazonec2-resources-for-iam-policies}{Resources Defined by Amazon EC2}.}

\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{DeviceId}{[required] The ID of the device.}

\item{LinkId}{The ID of the link.}
}
\description{
Associates a customer gateway with a device and optionally, with a link.
If you specify a link, it must be associated with the specified device.
}
\details{
You can only associate customer gateways that are connected to a VPN
attachment on a transit gateway. The transit gateway must be registered
in your global network. When you register a transit gateway, customer
gateways that are connected to the transit gateway are automatically
included in the global network. To list customer gateways that are
connected to a transit gateway, use the
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVpnConnections.html}{DescribeVpnConnections}
EC2 API and filter by \code{transit-gateway-id}.

You cannot associate a customer gateway with more than one device and
link.
}
\section{Request syntax}{
\preformatted{svc$associate_customer_gateway(
  CustomerGatewayArn = "string",
  GlobalNetworkId = "string",
  DeviceId = "string",
  LinkId = "string"
)
}
}

\keyword{internal}
