% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_associate_transit_gateway_connect_peer}
\alias{networkmanager_associate_transit_gateway_connect_peer}
\title{Associates a transit gateway Connect peer with a device, and optionally,
with a link}
\usage{
networkmanager_associate_transit_gateway_connect_peer(GlobalNetworkId,
  TransitGatewayConnectPeerArn, DeviceId, LinkId)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{TransitGatewayConnectPeerArn}{[required] The Amazon Resource Name (ARN) of the Connect peer.}

\item{DeviceId}{[required] The ID of the device.}

\item{LinkId}{The ID of the link.}
}
\description{
Associates a transit gateway Connect peer with a device, and optionally,
with a link. If you specify a link, it must be associated with the
specified device.
}
\details{
You can only associate transit gateway Connect peers that have been
created on a transit gateway that's registered in your global network.

You cannot associate a transit gateway Connect peer with more than one
device and link.
}
\section{Request syntax}{
\preformatted{svc$associate_transit_gateway_connect_peer(
  GlobalNetworkId = "string",
  TransitGatewayConnectPeerArn = "string",
  DeviceId = "string",
  LinkId = "string"
)
}
}

\keyword{internal}
