% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_global_network}
\alias{networkmanager_create_global_network}
\title{Creates a new, empty global network}
\usage{
networkmanager_create_global_network(Description, Tags)
}
\arguments{
\item{Description}{A description of the global network.

Length Constraints: Maximum length of 256 characters.}

\item{Tags}{The tags to apply to the resource during creation.}
}
\description{
Creates a new, empty global network.
}
\section{Request syntax}{
\preformatted{svc$create_global_network(
  Description = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
