% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_link}
\alias{networkmanager_create_link}
\title{Creates a new link for a specified site}
\usage{
networkmanager_create_link(GlobalNetworkId, Description, Type,
  Bandwidth, Provider, SiteId, Tags)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{Description}{A description of the link.

Length Constraints: Maximum length of 256 characters.}

\item{Type}{The type of the link.

Constraints: Cannot include the following characters: | \\ ^

Length Constraints: Maximum length of 128 characters.}

\item{Bandwidth}{[required] The upload speed and download speed in Mbps.}

\item{Provider}{The provider of the link.

Constraints: Cannot include the following characters: | \\ ^

Length Constraints: Maximum length of 128 characters.}

\item{SiteId}{[required] The ID of the site.}

\item{Tags}{The tags to apply to the resource during creation.}
}
\description{
Creates a new link for a specified site.
}
\section{Request syntax}{
\preformatted{svc$create_link(
  GlobalNetworkId = "string",
  Description = "string",
  Type = "string",
  Bandwidth = list(
    UploadSpeed = 123,
    DownloadSpeed = 123
  ),
  Provider = "string",
  SiteId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
