% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_delete_global_network}
\alias{networkmanager_delete_global_network}
\title{Deletes an existing global network}
\usage{
networkmanager_delete_global_network(GlobalNetworkId)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}
}
\description{
Deletes an existing global network. You must first delete all global
network objects (devices, links, and sites) and deregister all transit
gateways.
}
\section{Request syntax}{
\preformatted{svc$delete_global_network(
  GlobalNetworkId = "string"
)
}
}

\keyword{internal}
