% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_disassociate_customer_gateway}
\alias{networkmanager_disassociate_customer_gateway}
\title{Disassociates a customer gateway from a device and a link}
\usage{
networkmanager_disassociate_customer_gateway(GlobalNetworkId,
  CustomerGatewayArn)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{CustomerGatewayArn}{[required] The Amazon Resource Name (ARN) of the customer gateway. For more
information, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/list_amazonec2.html}{Resources Defined by Amazon EC2}.}
}
\description{
Disassociates a customer gateway from a device and a link.
}
\section{Request syntax}{
\preformatted{svc$disassociate_customer_gateway(
  GlobalNetworkId = "string",
  CustomerGatewayArn = "string"
)
}
}

\keyword{internal}
