% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_connections}
\alias{networkmanager_get_connections}
\title{Gets information about one or more of your connections in a global
network}
\usage{
networkmanager_get_connections(GlobalNetworkId, ConnectionIds, DeviceId,
  MaxResults, NextToken)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{ConnectionIds}{One or more connection IDs.}

\item{DeviceId}{The ID of the device.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Gets information about one or more of your connections in a global
network.
}
\section{Request syntax}{
\preformatted{svc$get_connections(
  GlobalNetworkId = "string",
  ConnectionIds = list(
    "string"
  ),
  DeviceId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
