% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_customer_gateway_associations}
\alias{networkmanager_get_customer_gateway_associations}
\title{Gets the association information for customer gateways that are
associated with devices and links in your global network}
\usage{
networkmanager_get_customer_gateway_associations(GlobalNetworkId,
  CustomerGatewayArns, MaxResults, NextToken)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{CustomerGatewayArns}{One or more customer gateway Amazon Resource Names (ARNs). For more
information, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/list_amazonec2.html}{Resources Defined by Amazon EC2}.
The maximum is 10.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Gets the association information for customer gateways that are
associated with devices and links in your global network.
}
\section{Request syntax}{
\preformatted{svc$get_customer_gateway_associations(
  GlobalNetworkId = "string",
  CustomerGatewayArns = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
