% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_transit_gateway_connect_peer_associations}
\alias{networkmanager_get_transit_gateway_connect_peer_associations}
\title{Gets information about one or more of your transit gateway Connect peer
associations in a global network}
\usage{
networkmanager_get_transit_gateway_connect_peer_associations(
  GlobalNetworkId, TransitGatewayConnectPeerArns, MaxResults, NextToken)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{TransitGatewayConnectPeerArns}{One or more transit gateway Connect peer Amazon Resource Names (ARNs).}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Gets information about one or more of your transit gateway Connect peer
associations in a global network.
}
\section{Request syntax}{
\preformatted{svc$get_transit_gateway_connect_peer_associations(
  GlobalNetworkId = "string",
  TransitGatewayConnectPeerArns = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
