% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_register_transit_gateway}
\alias{networkmanager_register_transit_gateway}
\title{Registers a transit gateway in your global network}
\usage{
networkmanager_register_transit_gateway(GlobalNetworkId,
  TransitGatewayArn)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{TransitGatewayArn}{[required] The Amazon Resource Name (ARN) of the transit gateway. For more
information, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/list_amazonec2.html}{Resources Defined by Amazon EC2}.}
}
\description{
Registers a transit gateway in your global network. The transit gateway
can be in any AWS Region, but it must be owned by the same AWS account
that owns the global network. You cannot register a transit gateway in
more than one global network.
}
\section{Request syntax}{
\preformatted{svc$register_transit_gateway(
  GlobalNetworkId = "string",
  TransitGatewayArn = "string"
)
}
}

\keyword{internal}
