% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_update_device}
\alias{networkmanager_update_device}
\title{Updates the details for an existing device}
\usage{
networkmanager_update_device(GlobalNetworkId, DeviceId, Description,
  Type, Vendor, Model, SerialNumber, Location, SiteId)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{DeviceId}{[required] The ID of the device.}

\item{Description}{A description of the device.

Length Constraints: Maximum length of 256 characters.}

\item{Type}{The type of the device.}

\item{Vendor}{The vendor of the device.

Length Constraints: Maximum length of 128 characters.}

\item{Model}{The model of the device.

Length Constraints: Maximum length of 128 characters.}

\item{SerialNumber}{The serial number of the device.

Length Constraints: Maximum length of 128 characters.}

\item{Location}{}

\item{SiteId}{The ID of the site.}
}
\description{
Updates the details for an existing device. To remove information for
any of the parameters, specify an empty string.
}
\section{Request syntax}{
\preformatted{svc$update_device(
  GlobalNetworkId = "string",
  DeviceId = "string",
  Description = "string",
  Type = "string",
  Vendor = "string",
  Model = "string",
  SerialNumber = "string",
  Location = list(
    Address = "string",
    Latitude = "string",
    Longitude = "string"
  ),
  SiteId = "string"
)
}
}

\keyword{internal}
