% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_delete_user_profile}
\alias{opsworks_delete_user_profile}
\title{Deletes a user profile}
\usage{
opsworks_delete_user_profile(IamUserArn)
}
\arguments{
\item{IamUserArn}{[required] The user's IAM ARN. This can also be a federated user's ARN.}
}
\description{
Deletes a user profile.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have an
attached policy that explicitly grants permissions. For more information
about user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{
\preformatted{opsworks$delete_user_profile(
  IamUserArn = "string"
)
}
}

\keyword{internal}
