% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_deregister_instance}
\alias{opsworks_deregister_instance}
\title{Deregister a registered Amazon EC2 or on-premises instance}
\usage{
opsworks_deregister_instance(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The instance ID.}
}
\description{
Deregister a registered Amazon EC2 or on-premises instance. This action
removes the instance from the stack and returns it to your control. This
action cannot be used with instances that were created with AWS OpsWorks
Stacks.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a
Manage permissions level for the stack or an attached policy that
explicitly grants permissions. For more information on user permissions,
see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{
\preformatted{svc$deregister_instance(
  InstanceId = "string"
)
}
}

\keyword{internal}
