% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_stack_provisioning_parameters}
\alias{opsworks_describe_stack_provisioning_parameters}
\title{Requests a description of a stack's provisioning parameters}
\usage{
opsworks_describe_stack_provisioning_parameters(StackId)
}
\arguments{
\item{StackId}{[required] The stack ID.}
}
\description{
Requests a description of a stack\'s provisioning parameters.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a
Show, Deploy, or Manage permissions level for the stack or an attached
policy that explicitly grants permissions. For more information about
user permissions, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{
\preformatted{svc$describe_stack_provisioning_parameters(
  StackId = "string"
)
}
}

\keyword{internal}
