% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_grant_access}
\alias{opsworks_grant_access}
\title{This action can be used only with Windows stacks}
\usage{
opsworks_grant_access(InstanceId, ValidForInMinutes)
}
\arguments{
\item{InstanceId}{[required] The instance's AWS OpsWorks Stacks ID.}

\item{ValidForInMinutes}{The length of time (in minutes) that the grant is valid. When the grant
expires at the end of this period, the user will no longer be able to
use the credentials to log in. If the user is logged in at the time, he
or she automatically will be logged out.}
}
\description{
This action can be used only with Windows stacks.
}
\details{
Grants RDP access to a Windows instance for a specified time period.
}
\section{Request syntax}{
\preformatted{opsworks$grant_access(
  InstanceId = "string",
  ValidForInMinutes = 123
)
}
}

\keyword{internal}
