% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_tags_for_resource}
\alias{organizations_list_tags_for_resource}
\title{Lists tags for the specified resource}
\usage{
organizations_list_tags_for_resource(ResourceId, NextToken)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource that you want to retrieve tags for.}

\item{NextToken}{Use this parameter if you receive a \code{NextToken} response in a previous
request that indicates that there is more output available. Set it to
the value of the previous call's \code{NextToken} response to indicate where
the output should continue from.}
}
\description{
Lists tags for the specified resource.
}
\details{
Currently, you can list tags on an account in AWS Organizations.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceId = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
