% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_tag_resource}
\alias{organizations_tag_resource}
\title{Adds one or more tags to the specified resource}
\usage{
organizations_tag_resource(ResourceId, Tags)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource to add a tag to.}

\item{Tags}{[required] The tag to add to the specified resource. Specifying the tag key is
required. You can set the value of a tag to an empty string, but you
can't set the value of a tag to null.}
}
\description{
Adds one or more tags to the specified resource.
}
\details{
Currently, you can tag and untag accounts in AWS Organizations.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
