% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_untag_resource}
\alias{organizations_untag_resource}
\title{Removes a tag from the specified resource}
\usage{
organizations_untag_resource(ResourceId, TagKeys)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource to remove the tag from.}

\item{TagKeys}{[required] The tag to remove from the specified resource.}
}
\description{
Removes a tag from the specified resource.
}
\details{
Currently, you can tag and untag accounts in AWS Organizations.

This operation can be called only from the organization's master
account.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceId = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
