% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outposts_service.R
\name{outposts}
\alias{outposts}
\title{AWS Outposts}
\usage{
outposts(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
AWS Outposts is a fully-managed service that extends AWS infrastructure,
APIs, and tools to customer premises. By providing local access to
AWS-managed infrastructure, AWS Outposts enables customers to build and
run applications on premises using the same programming interfaces as in
AWS Regions, while using local compute and storage resources for lower
latency and local data processing needs.
}
\section{Service syntax}{
\preformatted{svc <- outposts(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=outposts_create_outpost]{create_outpost} \tab Creates an Outpost \cr
\link[=outposts_delete_outpost]{delete_outpost} \tab Deletes the Outpost \cr
\link[=outposts_delete_site]{delete_site} \tab Deletes the site \cr
\link[=outposts_get_outpost]{get_outpost} \tab Gets information about the specified Outpost \cr
\link[=outposts_get_outpost_instance_types]{get_outpost_instance_types} \tab Lists the instance types for the specified Outpost\cr
\link[=outposts_list_outposts]{list_outposts} \tab List the Outposts for your AWS account \cr
\link[=outposts_list_sites]{list_sites} \tab Lists the sites for the specified AWS account
}
}

\examples{
\dontrun{
svc <- outposts()
svc$create_outpost(
  Foo = 123
)
}

}
