% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outposts_operations.R
\name{outposts_delete_outpost}
\alias{outposts_delete_outpost}
\title{Deletes the Outpost}
\usage{
outposts_delete_outpost(OutpostId)
}
\arguments{
\item{OutpostId}{[required]}
}
\description{
Deletes the Outpost.
}
\section{Request syntax}{
\preformatted{svc$delete_outpost(
  OutpostId = "string"
)
}
}

\keyword{internal}
