% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outposts_operations.R
\name{outposts_delete_site}
\alias{outposts_delete_site}
\title{Deletes the site}
\usage{
outposts_delete_site(SiteId)
}
\arguments{
\item{SiteId}{[required]}
}
\description{
Deletes the site.
}
\section{Request syntax}{
\preformatted{svc$delete_site(
  SiteId = "string"
)
}
}

\keyword{internal}
