% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outposts_operations.R
\name{outposts_get_outpost}
\alias{outposts_get_outpost}
\title{Gets information about the specified Outpost}
\usage{
outposts_get_outpost(OutpostId)
}
\arguments{
\item{OutpostId}{[required]}
}
\description{
Gets information about the specified Outpost.
}
\section{Request syntax}{
\preformatted{svc$get_outpost(
  OutpostId = "string"
)
}
}

\keyword{internal}
