% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outposts_operations.R
\name{outposts_get_outpost_instance_types}
\alias{outposts_get_outpost_instance_types}
\title{Lists the instance types for the specified Outpost}
\usage{
outposts_get_outpost_instance_types(OutpostId, NextToken, MaxResults)
}
\arguments{
\item{OutpostId}{[required]}

\item{NextToken}{}

\item{MaxResults}{}
}
\description{
Lists the instance types for the specified Outpost.
}
\section{Request syntax}{
\preformatted{svc$get_outpost_instance_types(
  OutpostId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
