% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outposts_operations.R
\name{outposts_list_outposts}
\alias{outposts_list_outposts}
\title{List the Outposts for your AWS account}
\usage{
outposts_list_outposts(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{}

\item{MaxResults}{}
}
\description{
List the Outposts for your AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_outposts(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
