% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outposts_operations.R
\name{outposts_list_sites}
\alias{outposts_list_sites}
\title{Lists the sites for the specified AWS account}
\usage{
outposts_list_sites(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{}

\item{MaxResults}{}
}
\description{
Lists the sites for the specified AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_sites(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
