% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outposts_operations.R
\name{outposts_untag_resource}
\alias{outposts_untag_resource}
\title{Removes tags from the specified resource}
\usage{
outposts_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagKeys}{[required] The tag keys.}
}
\description{
Removes tags from the specified resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
