% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{pain_class}
\alias{pain_class}
\title{Identify pain class based on pain score}
\usage{
pain_class(score, type = c("binary", "trinary"))
}
\arguments{
\item{score}{pain scores generated by \code{\link{pain_score}}}

\item{type}{binary or trinary (ternary) pain classifications}
}
\value{
one or more pain classes
}
\description{
Identify pain class based on pain score
}
