% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{pain_score}
\alias{pain_score}
\title{Scoring pain from paw features}
\usage{
pain_score(
  paw.features,
  strains = NULL,
  strain.standards = jones2020.standards,
  feature.set = c("post.peak", "pre.peak")
)
}
\arguments{
\item{paw.features}{the paw features returned by
\code{\link{extract_features}}}

\item{strains}{a vector of containing strain information for each mouse}

\item{strain.standards}{z-scores (centering and scaling) information
by strain}

\item{feature.set}{use either pre-peak or post-peak features}
}
\value{
one or more pain scores as a vector
}
\description{
Returns a pain score based on Jones et al. 2020
}
