% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws_package.R
\docType{package}
\name{paws-package}
\alias{paws}
\alias{paws-package}
\title{paws: Paws Amazon Web Services SDK}
\description{
An Amazon Web Services software development kit.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
