% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\docType{package}
\name{paws}
\alias{paws}
\title{paws: Pain Assessment at Withdrawal Speeds}
\description{
Automated pain scoring from paw withdrawl tracking data based on Jones et
al. (2020) A machine-vision approach for automated pain measurement at
millisecond timescales. This R package takes paw trajectory data in response
to a stimulus and provides an automated scoring of pain.
}
