% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{paws_dashboard}
\alias{paws_dashboard}
\title{Launch the PAWS Dashboard}
\usage{
paws_dashboard()
}
\value{
The PAWS dashboard.
}
\description{
This function launches a Shiny dashboard for running PAWS interactively without
any code. Settings will be able to be specified in the Dashboard and visualizations
that are the output of many PAWS functions will appear in their corresponding tabs here, too.
}
