% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawscore-package.R
\docType{package}
\name{pawscore-package}
\alias{pawscore}
\alias{pawscore-package}
\title{pawscore: Pain Assessment at Withdrawal Speeds (PAWS)}
\description{
Automated pain scoring from paw withdrawal tracking data. Based on Jones et al. (2020) "A machine-vision approach for automated pain measurement at millisecond timescales" \doi{10.7554/eLife.57258}.
}
\author{
\strong{Maintainer}: Colin Twomey \email{crtwomey@gmail.com}

Authors:
\itemize{
  \item Jessica Jones
  \item William Foster
  \item Joshua Plotkin
  \item Ishmail Abdus-Saboor
}

}
\keyword{internal}
