% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_schema}
\alias{personalize_delete_schema}
\title{Deletes a schema}
\usage{
personalize_delete_schema(schemaArn)
}
\arguments{
\item{schemaArn}{[required] The Amazon Resource Name (ARN) of the schema to delete.}
}
\description{
Deletes a schema. Before deleting a schema, you must delete all datasets
referencing the schema. For more information on schemas, see
CreateSchema.
}
\section{Request syntax}{
\preformatted{svc$delete_schema(
  schemaArn = "string"
)
}
}

\keyword{internal}
