% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_dataset}
\alias{personalize_describe_dataset}
\title{Describes the given dataset}
\usage{
personalize_describe_dataset(datasetArn)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Name (ARN) of the dataset to describe.}
}
\description{
Describes the given dataset. For more information on datasets, see
\code{\link[=personalize_create_dataset]{create_dataset}}.
}
\section{Request syntax}{
\preformatted{svc$describe_dataset(
  datasetArn = "string"
)
}
}

\keyword{internal}
