% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_solution}
\alias{personalize_describe_solution}
\title{Describes a solution}
\usage{
personalize_describe_solution(solutionArn)
}
\arguments{
\item{solutionArn}{[required] The Amazon Resource Name (ARN) of the solution to describe.}
}
\description{
Describes a solution. For more information on solutions, see
\code{\link[=personalize_create_solution]{create_solution}}.
}
\section{Request syntax}{
\preformatted{svc$describe_solution(
  solutionArn = "string"
)
}
}

\keyword{internal}
