% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_solution_version}
\alias{personalize_describe_solution_version}
\title{Describes a specific version of a solution}
\usage{
personalize_describe_solution_version(solutionVersionArn)
}
\arguments{
\item{solutionVersionArn}{[required] The Amazon Resource Name (ARN) of the solution version.}
}
\description{
Describes a specific version of a solution. For more information on
solutions, see \code{\link[=personalize_create_solution]{create_solution}}.
}
\section{Request syntax}{
\preformatted{svc$describe_solution_version(
  solutionVersionArn = "string"
)
}
}

\keyword{internal}
