% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_recipes}
\alias{personalize_list_recipes}
\title{Returns a list of available recipes}
\usage{
personalize_list_recipes(recipeProvider, nextToken, maxResults)
}
\arguments{
\item{recipeProvider}{The default is \code{SERVICE}.}

\item{nextToken}{A token returned from the previous call to \code{ListRecipes} for getting the
next set of recipes (if they exist).}

\item{maxResults}{The maximum number of recipes to return.}
}
\description{
Returns a list of available recipes. The response provides the
properties for each recipe, including the recipe's Amazon Resource Name
(ARN).
}
\section{Request syntax}{
\preformatted{svc$list_recipes(
  recipeProvider = "SERVICE",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
