% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_schemas}
\alias{personalize_list_schemas}
\title{Returns the list of schemas associated with the account}
\usage{
personalize_list_schemas(nextToken, maxResults)
}
\arguments{
\item{nextToken}{A token returned from the previous call to \code{ListSchemas} for getting the
next set of schemas (if they exist).}

\item{maxResults}{The maximum number of schemas to return.}
}
\description{
Returns the list of schemas associated with the account. The response
provides the properties for each schema, including the Amazon Resource
Name (ARN). For more information on schemas, see CreateSchema.
}
\section{Request syntax}{
\preformatted{svc$list_schemas(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
