% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalizeevents_operations.R
\name{personalizeevents_put_users}
\alias{personalizeevents_put_users}
\title{Adds one or more users to a Users dataset}
\usage{
personalizeevents_put_users(datasetArn, users)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Number (ARN) of the Users dataset you are adding the
user or users to.}

\item{users}{[required] A list of user data.}
}
\description{
Adds one or more users to a Users dataset. For more information see
importing-users.
}
\section{Request syntax}{
\preformatted{svc$put_users(
  datasetArn = "string",
  users = list(
    list(
      userId = "string",
      properties = "string"
    )
  )
)
}
}

\keyword{internal}
