% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_import_job}
\alias{pinpoint_create_import_job}
\title{Creates or updates an import job}
\usage{
pinpoint_create_import_job(ApplicationId, ImportJobRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{ImportJobRequest}{[required]}
}
\description{
Creates or updates an import job.
}
\section{Request syntax}{
\preformatted{pinpoint$create_import_job(
  ApplicationId = "string",
  ImportJobRequest = list(
    DefineSegment = TRUE|FALSE,
    ExternalId = "string",
    Format = "CSV"|"JSON",
    RegisterEndpoints = TRUE|FALSE,
    RoleArn = "string",
    S3Url = "string",
    SegmentId = "string",
    SegmentName = "string"
  )
)
}
}

\keyword{internal}
