% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_sms_template}
\alias{pinpoint_create_sms_template}
\title{Creates a message template for messages that are sent through the SMS
channel}
\usage{
pinpoint_create_sms_template(SMSTemplateRequest, TemplateName)
}
\arguments{
\item{SMSTemplateRequest}{[required]}

\item{TemplateName}{[required] The name of the message template. A template name must start with an
alphanumeric character and can contain a maximum of 128 characters. The
characters can be alphanumeric characters, underscores (\\_), or hyphens
(-). Template names are case sensitive.}
}
\description{
Creates a message template for messages that are sent through the SMS
channel.
}
\section{Request syntax}{
\preformatted{svc$create_sms_template(
  SMSTemplateRequest = list(
    Body = "string",
    DefaultSubstitutions = "string",
    tags = list(
      "string"
    ),
    TemplateDescription = "string"
  ),
  TemplateName = "string"
)
}
}

\keyword{internal}
