% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_endpoint}
\alias{pinpoint_delete_endpoint}
\title{Deletes an endpoint}
\usage{
pinpoint_delete_endpoint(ApplicationId, EndpointId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{EndpointId}{[required] The unique ID of the endpoint.}
}
\description{
Deletes an endpoint.
}
\section{Request syntax}{
\preformatted{svc$delete_endpoint(
  ApplicationId = "string",
  EndpointId = "string"
)
}
}

\keyword{internal}
