% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_segment}
\alias{pinpoint_delete_segment}
\title{Deletes a segment}
\usage{
pinpoint_delete_segment(ApplicationId, SegmentId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{SegmentId}{[required] The unique ID of the segment.}
}
\description{
Deletes a segment.
}
\section{Request syntax}{
\preformatted{pinpoint$delete_segment(
  ApplicationId = "string",
  SegmentId = "string"
)
}
}

\keyword{internal}
