% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_adm_channel}
\alias{pinpoint_get_adm_channel}
\title{Get an ADM channel}
\usage{
pinpoint_get_adm_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get an ADM channel.
}
\section{Request syntax}{
\preformatted{svc$get_adm_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
