% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_apns_voip_channel}
\alias{pinpoint_get_apns_voip_channel}
\title{Get an APNS VoIP channel}
\usage{
pinpoint_get_apns_voip_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get an APNS VoIP channel
}
\section{Request syntax}{
\preformatted{pinpoint$get_apns_voip_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
