% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_apns_voip_sandbox_channel}
\alias{pinpoint_get_apns_voip_sandbox_channel}
\title{Get an APNS VoIPSandbox channel}
\usage{
pinpoint_get_apns_voip_sandbox_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get an APNS VoIPSandbox channel
}
\section{Request syntax}{
\preformatted{svc$get_apns_voip_sandbox_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
