% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_app}
\alias{pinpoint_get_app}
\title{Retrieves information about an application}
\usage{
pinpoint_get_app(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Retrieves information about an application.
}
\section{Request syntax}{
\preformatted{svc$get_app(
  ApplicationId = "string"
)
}
}

\keyword{internal}
