% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_apps}
\alias{pinpoint_get_apps}
\title{Returns information about your apps}
\usage{
pinpoint_get_apps(PageSize, Token)
}
\arguments{
\item{PageSize}{The number of entries you want on each page in the response.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns information about your apps.
}
\section{Request syntax}{
\preformatted{svc$get_apps(
  PageSize = "string",
  Token = "string"
)
}
}

\keyword{internal}
