% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_campaign_version}
\alias{pinpoint_get_campaign_version}
\title{Returns information about a specific version of a campaign}
\usage{
pinpoint_get_campaign_version(ApplicationId, CampaignId, Version)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{CampaignId}{[required] The unique ID of the campaign.}

\item{Version}{[required] The version of the campaign.}
}
\description{
Returns information about a specific version of a campaign.
}
\section{Request syntax}{
\preformatted{pinpoint$get_campaign_version(
  ApplicationId = "string",
  CampaignId = "string",
  Version = "string"
)
}
}

\keyword{internal}
