% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_endpoint}
\alias{pinpoint_get_endpoint}
\title{Returns information about an endpoint}
\usage{
pinpoint_get_endpoint(ApplicationId, EndpointId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{EndpointId}{[required] The unique ID of the endpoint.}
}
\description{
Returns information about an endpoint.
}
\section{Request syntax}{
\preformatted{pinpoint$get_endpoint(
  ApplicationId = "string",
  EndpointId = "string"
)
}
}

\keyword{internal}
