% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_event_stream}
\alias{pinpoint_get_event_stream}
\title{Returns the event stream for an app}
\usage{
pinpoint_get_event_stream(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Returns the event stream for an app.
}
\section{Request syntax}{
\preformatted{pinpoint$get_event_stream(
  ApplicationId = "string"
)
}
}

\keyword{internal}
