% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_gcm_channel}
\alias{pinpoint_get_gcm_channel}
\title{Returns information about the GCM channel for an app}
\usage{
pinpoint_get_gcm_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Returns information about the GCM channel for an app.
}
\section{Request syntax}{
\preformatted{pinpoint$get_gcm_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
