% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_segment}
\alias{pinpoint_get_segment}
\title{Returns information about a segment}
\usage{
pinpoint_get_segment(ApplicationId, SegmentId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{SegmentId}{[required] The unique ID of the segment.}
}
\description{
Returns information about a segment.
}
\section{Request syntax}{
\preformatted{svc$get_segment(
  ApplicationId = "string",
  SegmentId = "string"
)
}
}

\keyword{internal}
