% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_segment_version}
\alias{pinpoint_get_segment_version}
\title{Returns information about a segment version}
\usage{
pinpoint_get_segment_version(ApplicationId, SegmentId, Version)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{SegmentId}{[required] The unique ID of the segment.}

\item{Version}{[required] The segment version.}
}
\description{
Returns information about a segment version.
}
\section{Request syntax}{
\preformatted{pinpoint$get_segment_version(
  ApplicationId = "string",
  SegmentId = "string",
  Version = "string"
)
}
}

\keyword{internal}
