% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_segments}
\alias{pinpoint_get_segments}
\title{Used to get information about your segments}
\usage{
pinpoint_get_segments(ApplicationId, PageSize, Token)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{PageSize}{The number of entries you want on each page in the response.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Used to get information about your segments.
}
\section{Request syntax}{
\preformatted{pinpoint$get_segments(
  ApplicationId = "string",
  PageSize = "string",
  Token = "string"
)
}
}

\keyword{internal}
