% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_list_tags_for_resource}
\alias{pinpoint_list_tags_for_resource}
\title{List tags for resource}
\usage{
pinpoint_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required]}
}
\description{
List tags for resource
}
\section{Request syntax}{
\preformatted{pinpoint$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
