% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_phone_number_validate}
\alias{pinpoint_phone_number_validate}
\title{Retrieves information about a phone number}
\usage{
pinpoint_phone_number_validate(NumberValidateRequest)
}
\arguments{
\item{NumberValidateRequest}{[required]}
}
\description{
Retrieves information about a phone number.
}
\section{Request syntax}{
\preformatted{svc$phone_number_validate(
  NumberValidateRequest = list(
    IsoCountryCode = "string",
    PhoneNumber = "string"
  )
)
}
}

\keyword{internal}
