% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_adm_channel}
\alias{pinpoint_update_adm_channel}
\title{Updates the ADM channel settings for an application}
\usage{
pinpoint_update_adm_channel(ADMChannelRequest, ApplicationId)
}
\arguments{
\item{ADMChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Updates the ADM channel settings for an application.
}
\section{Request syntax}{
\preformatted{svc$update_adm_channel(
  ADMChannelRequest = list(
    ClientId = "string",
    ClientSecret = "string",
    Enabled = TRUE|FALSE
  ),
  ApplicationId = "string"
)
}
}

\keyword{internal}
