% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_email_template}
\alias{pinpoint_update_email_template}
\title{Updates an existing message template that you can use in messages that
are sent through the email channel}
\usage{
pinpoint_update_email_template(EmailTemplateRequest, TemplateName)
}
\arguments{
\item{EmailTemplateRequest}{[required]}

\item{TemplateName}{[required] The name of the message template. A template name must start with an
alphanumeric character and can contain a maximum of 128 characters. The
characters can be alphanumeric characters, underscores (\\_), or hyphens
(-). Template names are case sensitive.}
}
\description{
Updates an existing message template that you can use in messages that
are sent through the email channel.
}
\section{Request syntax}{
\preformatted{svc$update_email_template(
  EmailTemplateRequest = list(
    DefaultSubstitutions = "string",
    HtmlPart = "string",
    Subject = "string",
    tags = list(
      "string"
    ),
    TemplateDescription = "string",
    TextPart = "string"
  ),
  TemplateName = "string"
)
}
}

\keyword{internal}
