% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_gcm_channel}
\alias{pinpoint_update_gcm_channel}
\title{Enables the GCM channel for an application or updates the status and
settings of the GCM channel for an application}
\usage{
pinpoint_update_gcm_channel(ApplicationId, GCMChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{GCMChannelRequest}{[required]}
}
\description{
Enables the GCM channel for an application or updates the status and
settings of the GCM channel for an application.
}
\section{Request syntax}{
\preformatted{svc$update_gcm_channel(
  ApplicationId = "string",
  GCMChannelRequest = list(
    ApiKey = "string",
    Enabled = TRUE|FALSE
  )
)
}
}

\keyword{internal}
