% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_sms_channel}
\alias{pinpoint_update_sms_channel}
\title{Update an SMS channel}
\usage{
pinpoint_update_sms_channel(ApplicationId, SMSChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{SMSChannelRequest}{[required]}
}
\description{
Update an SMS channel.
}
\section{Request syntax}{
\preformatted{svc$update_sms_channel(
  ApplicationId = "string",
  SMSChannelRequest = list(
    Enabled = TRUE|FALSE,
    SenderId = "string",
    ShortCode = "string"
  )
)
}
}

\keyword{internal}
