% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_voice_channel}
\alias{pinpoint_update_voice_channel}
\title{Updates the status and settings of the voice channel for an application}
\usage{
pinpoint_update_voice_channel(ApplicationId, VoiceChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{VoiceChannelRequest}{[required]}
}
\description{
Updates the status and settings of the voice channel for an application.
}
\section{Request syntax}{
\preformatted{svc$update_voice_channel(
  ApplicationId = "string",
  VoiceChannelRequest = list(
    Enabled = TRUE|FALSE
  )
)
}
}

\keyword{internal}
