% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_delete_email_identity}
\alias{pinpointemail_delete_email_identity}
\title{Deletes an email identity that you previously verified for use with
Amazon Pinpoint}
\usage{
pinpointemail_delete_email_identity(EmailIdentity)
}
\arguments{
\item{EmailIdentity}{[required] The identity (that is, the email address or domain) that you want to
delete from your Amazon Pinpoint account.}
}
\description{
Deletes an email identity that you previously verified for use with
Amazon Pinpoint. An identity can be either an email address or a domain
name.
}
\section{Request syntax}{
\preformatted{pinpointemail$delete_email_identity(
  EmailIdentity = "string"
)
}
}

\keyword{internal}
