% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_get_deliverability_test_report}
\alias{pinpointemail_get_deliverability_test_report}
\title{Retrieve the results of a predictive inbox placement test}
\usage{
pinpointemail_get_deliverability_test_report(ReportId)
}
\arguments{
\item{ReportId}{[required] A unique string that identifies the predictive inbox placement test.}
}
\description{
Retrieve the results of a predictive inbox placement test.
}
\section{Request syntax}{
\preformatted{svc$get_deliverability_test_report(
  ReportId = "string"
)
}
}

\keyword{internal}
