% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_untag_resource}
\alias{pinpointemail_untag_resource}
\title{Remove one or more tags (keys and values) from a specified resource}
\usage{
pinpointemail_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to remove
one or more tags from.}

\item{TagKeys}{[required] The tags (tag keys) that you want to remove from the resource. When you
specify a tag key, the action removes both that key and its associated
tag value.

To remove more than one tag from the resource, append the \code{TagKeys}
parameter and argument for each additional tag to remove, separated by
an ampersand. For example:
\code{/v1/email/tags?ResourceArn=ResourceArn&amp;TagKeys=Key1&amp;TagKeys=Key2}}
}
\description{
Remove one or more tags (keys and values) from a specified resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
