% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filter_plots.R
\name{plot_filter_graphs}
\alias{plot_filter_graphs}
\title{Plot a given trajectory under all available filters}
\usage{
plot_filter_graphs(
  csv_or_path,
  p_cutoff,
  reference_distance = NA,
  manual_scale_factor = NA,
  fps = 2000,
  fixed_baseline = 1,
  y_threshold = 0.1,
  savgol_window_length = 25,
  savgol_filter_smoothing_multiplier = 3,
  median_window_length = 25,
  average_window_length = 25,
  body_part = "center",
  axis = "y"
)
}
\arguments{
\item{csv_or_path}{Either a csv loaded as an R object, or the full path to that CSV.}

\item{p_cutoff}{The confidence value at which all tracked points below this
threshold are replaced by linear interpolation between two higher-confidence points.}

\item{reference_distance}{The 'real-world-length' between your reference objects (can be in mm, cm, etc).
If you indicate a \code{manual_scale_factor}, this setting will be overrided by that factor.}

\item{manual_scale_factor}{The 'real-world-length' to pixel conversion factor (i.e. millimeters/pixel).
If you are using two reference points, you can ignore this parameter.}

\item{fps}{The frames per second of your CSV.}

\item{fixed_baseline}{The height (in units of choice) above the lowest y-axis position of the paw, used
to determine the baseline for activity. the \code{y_threshold} parameter is used to set a baseline for this
level of activity.}

\item{y_threshold}{The threshold (in units of choice) above the fixed baseline at which the start and end
time-points of activity are determined.}

\item{savgol_window_length}{The rolling window length of Savitzky-Golay filter smoothing to apply
to your tracking trajectory.}

\item{median_window_length}{The rolling window length of median filter smoothing to apply
to your tracking trajectory.}

\item{average_window_length}{The rolling window length of average filter smoothing to apply
to your tracking trajectory.}

\item{body_part}{The body_part you wish to plot.}

\item{axis}{The axis you wish to plot.}
}
\value{
A figure with panels (each indicating the plotted track under a given filter
and level and smoothing).
}
\description{
This function loads a CSV containing tracking data. Given a particular body-part,
axis, and level of smoothing, the function plots the trajectory of the body-part
along that axis.
}
