% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_pain_scores.R
\name{plot_pain_scores}
\alias{plot_pain_scores}
\title{Plot pain score distributions}
\usage{
plot_pain_scores(csv_path, peak = "pre")
}
\arguments{
\item{csv_path}{The path to the aggregated CSV exported from the \code{paws_analysis} function}

\item{peak}{Whether to plot pre-peak (\code{pre}) or post-peak (\code{post}) pain scores.}
}
\value{
A density plot of pain score distributions.
}
\description{
This function plots the distribution of pain scores, at either pre-peak or post-peak.
}
