% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_pain_scores.R
\name{plot_paws_features}
\alias{plot_paws_features}
\title{Plot PAWS features}
\usage{
plot_paws_features(csv_path, plot_features)
}
\arguments{
\item{csv_path}{The path to the aggregated CSV exported from the \code{paws_analysis} function}

\item{plot_features}{The PAWS feature to plot. Options are \verb{pre-peak max height},
\verb{pre-peak max x-velocity}, \verb{pre-peak max y-velocity}, \verb{pre-peak distance traveled},
\verb{post-peak max height}, \verb{post-peak max x-velocity}, \verb{post-peak max y-velocity},
\verb{post-peak distance traveled}, \verb{post-peak number of shakes}, \verb{post-peak shaking duration},
\verb{post-peak guarding duration}}
}
\value{
A bar plot of PAWS features over a distribution.
}
\description{
This function plots a given PAWS feature across stimuli, groups, and body-parts.
}
