% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polly_operations.R
\name{polly_get_lexicon}
\alias{polly_get_lexicon}
\title{Returns the content of the specified pronunciation lexicon stored in an
AWS Region}
\usage{
polly_get_lexicon(Name)
}
\arguments{
\item{Name}{[required] Name of the lexicon.}
}
\description{
Returns the content of the specified pronunciation lexicon stored in an
AWS Region. For more information, see \href{https://docs.aws.amazon.com/polly/latest/dg/managing-lexicons.html}{Managing Lexicons}.
}
\section{Request syntax}{
\preformatted{svc$get_lexicon(
  Name = "string"
)
}
}

\section{Examples}{
\preformatted{# Returns the content of the specified pronunciation lexicon stored in an
# AWS Region.
svc$get_lexicon(
  Name = ""
)
}
}

\keyword{internal}
