% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polly_operations.R
\name{polly_list_lexicons}
\alias{polly_list_lexicons}
\title{Returns a list of pronunciation lexicons stored in an AWS Region}
\usage{
polly_list_lexicons(NextToken)
}
\arguments{
\item{NextToken}{An opaque pagination token returned from previous
\code{\link[=polly_list_lexicons]{list_lexicons}} operation. If present, indicates
where to continue the list of lexicons.}
}
\description{
Returns a list of pronunciation lexicons stored in an AWS Region. For
more information, see \href{https://docs.aws.amazon.com/polly/latest/dg/managing-lexicons.html}{Managing Lexicons}.
}
\section{Request syntax}{
\preformatted{svc$list_lexicons(
  NextToken = "string"
)
}
}

\examples{
\dontrun{
# Returns a list of pronunciation lexicons stored in an AWS Region.
svc$list_lexicons()
}

}
\keyword{internal}
