% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_update_workspace_alias}
\alias{prometheusservice_update_workspace_alias}
\title{Updates an AMP workspace alias}
\usage{
prometheusservice_update_workspace_alias(alias, clientToken,
  workspaceId)
}
\arguments{
\item{alias}{The new alias of the workspace.}

\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}

\item{workspaceId}{[required] The ID of the workspace being updated.}
}
\description{
Updates an AMP workspace alias.
}
\section{Request syntax}{
\preformatted{svc$update_workspace_alias(
  alias = "string",
  clientToken = "string",
  workspaceId = "string"
)
}
}

\keyword{internal}
