% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_cancel_journal_kinesis_stream}
\alias{qldb_cancel_journal_kinesis_stream}
\title{Ends a given Amazon QLDB journal stream}
\usage{
qldb_cancel_journal_kinesis_stream(LedgerName, StreamId)
}
\arguments{
\item{LedgerName}{[required] The name of the ledger.}

\item{StreamId}{[required] The unique ID that QLDB assigns to each QLDB journal stream.}
}
\description{
Ends a given Amazon QLDB journal stream. Before a stream can be
canceled, its current status must be \code{ACTIVE}.

You can't restart a stream after you cancel it. Canceled QLDB stream
resources are subject to a 7-day retention period, so they are
automatically deleted after this limit expires.
}
\section{Request syntax}{
\preformatted{svc$cancel_journal_kinesis_stream(
  LedgerName = "string",
  StreamId = "string"
)
}
}

\keyword{internal}
