% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_create_ledger}
\alias{qldb_create_ledger}
\title{Creates a new ledger in your AWS account}
\usage{
qldb_create_ledger(Name, Tags, PermissionsMode, DeletionProtection)
}
\arguments{
\item{Name}{[required] The name of the ledger that you want to create. The name must be unique
among all of your ledgers in the current AWS Region.

Naming constraints for ledger names are defined in \href{https://docs.aws.amazon.com/qldb/latest/developerguide/limits.html#limits.naming}{Quotas in Amazon QLDB}
in the \emph{Amazon QLDB Developer Guide}.}

\item{Tags}{The key-value pairs to add as tags to the ledger that you want to
create. Tag keys are case sensitive. Tag values are case sensitive and
can be null.}

\item{PermissionsMode}{[required] The permissions mode to assign to the ledger that you want to create.}

\item{DeletionProtection}{The flag that prevents a ledger from being deleted by any user. If not
provided on ledger creation, this feature is enabled (\code{true}) by
default.

If deletion protection is enabled, you must first disable it before you
can delete the ledger using the QLDB API or the AWS Command Line
Interface (AWS CLI). You can disable it by calling the
\code{\link[=qldb_update_ledger]{update_ledger}} operation to set the flag to
\code{false}. The QLDB console disables deletion protection for you when you
use it to delete a ledger.}
}
\description{
Creates a new ledger in your AWS account.
}
\section{Request syntax}{
\preformatted{svc$create_ledger(
  Name = "string",
  Tags = list(
    "string"
  ),
  PermissionsMode = "ALLOW_ALL",
  DeletionProtection = TRUE|FALSE
)
}
}

\keyword{internal}
