% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_delete_ledger}
\alias{qldb_delete_ledger}
\title{Deletes a ledger and all of its contents}
\usage{
qldb_delete_ledger(Name)
}
\arguments{
\item{Name}{[required] The name of the ledger that you want to delete.}
}
\description{
Deletes a ledger and all of its contents. This action is irreversible.

If deletion protection is enabled, you must first disable it before you
can delete the ledger using the QLDB API or the AWS Command Line
Interface (AWS CLI). You can disable it by calling the
\code{\link[=qldb_update_ledger]{update_ledger}} operation to set the flag to
\code{false}. The QLDB console disables deletion protection for you when you
use it to delete a ledger.
}
\section{Request syntax}{
\preformatted{svc$delete_ledger(
  Name = "string"
)
}
}

\keyword{internal}
