% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_describe_journal_s3_export}
\alias{qldb_describe_journal_s3_export}
\title{Returns information about a journal export job, including the ledger
name, export ID, when it was created, current status, and its start and
end time export parameters}
\usage{
qldb_describe_journal_s3_export(Name, ExportId)
}
\arguments{
\item{Name}{[required] The name of the ledger.}

\item{ExportId}{[required] The unique ID of the journal export job that you want to describe.}
}
\description{
Returns information about a journal export job, including the ledger
name, export ID, when it was created, current status, and its start and
end time export parameters.
}
\details{
If the export job with the given \code{ExportId} doesn\'t exist, then throws
\code{ResourceNotFoundException}.

If the ledger with the given \code{Name} doesn\'t exist, then throws
\code{ResourceNotFoundException}.
}
\section{Request syntax}{
\preformatted{svc$describe_journal_s3_export(
  Name = "string",
  ExportId = "string"
)
}
}

\keyword{internal}
