% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_export_journal_to_s3}
\alias{qldb_export_journal_to_s3}
\title{Exports journal contents within a date and time range from a ledger into
a specified Amazon Simple Storage Service (Amazon S3) bucket}
\usage{
qldb_export_journal_to_s3(Name, InclusiveStartTime, ExclusiveEndTime,
  S3ExportConfiguration, RoleArn)
}
\arguments{
\item{Name}{[required] The name of the ledger.}

\item{InclusiveStartTime}{[required] The inclusive start date and time for the range of journal contents that
you want to export.

The \code{InclusiveStartTime} must be in \verb{ISO 8601} date and time format and
in Universal Coordinated Time (UTC). For example: \verb{2019-06-13T21:36:34Z}

The \code{InclusiveStartTime} must be before \code{ExclusiveEndTime}.

If you provide an \code{InclusiveStartTime} that is before the ledger's
\code{CreationDateTime}, Amazon QLDB defaults it to the ledger's
\code{CreationDateTime}.}

\item{ExclusiveEndTime}{[required] The exclusive end date and time for the range of journal contents that
you want to export.

The \code{ExclusiveEndTime} must be in \verb{ISO 8601} date and time format and in
Universal Coordinated Time (UTC). For example: \verb{2019-06-13T21:36:34Z}

The \code{ExclusiveEndTime} must be less than or equal to the current UTC
date and time.}

\item{S3ExportConfiguration}{[required] The configuration settings of the Amazon S3 bucket destination for your
export request.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that grants QLDB
permissions for a journal export job to do the following:
\itemize{
\item Write objects into your Amazon Simple Storage Service (Amazon S3)
bucket.
\item (Optional) Use your customer master key (CMK) in AWS Key Management
Service (AWS KMS) for server-side encryption of your exported data.
}}
}
\description{
Exports journal contents within a date and time range from a ledger into
a specified Amazon Simple Storage Service (Amazon S3) bucket. The data
is written as files in Amazon Ion format.
}
\details{
If the ledger with the given \code{Name} doesn't exist, then throws
\code{ResourceNotFoundException}.

If the ledger with the given \code{Name} is in \code{CREATING} status, then throws
\code{ResourcePreconditionNotMetException}.

You can initiate up to two concurrent journal export requests for each
ledger. Beyond this limit, journal export requests throw
\code{LimitExceededException}.
}
\section{Request syntax}{
\preformatted{svc$export_journal_to_s3(
  Name = "string",
  InclusiveStartTime = as.POSIXct(
    "2015-01-01"
  ),
  ExclusiveEndTime = as.POSIXct(
    "2015-01-01"
  ),
  S3ExportConfiguration = list(
    Bucket = "string",
    Prefix = "string",
    EncryptionConfiguration = list(
      ObjectEncryptionType = "SSE_KMS"|"SSE_S3"|"NO_ENCRYPTION",
      KmsKeyArn = "string"
    )
  ),
  RoleArn = "string"
)
}
}

\keyword{internal}
