% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_get_block}
\alias{qldb_get_block}
\title{Returns a block object at a specified address in a journal}
\usage{
qldb_get_block(Name, BlockAddress, DigestTipAddress)
}
\arguments{
\item{Name}{[required] The name of the ledger.}

\item{BlockAddress}{[required] The location of the block that you want to request. An address is an
Amazon Ion structure that has two fields: \code{strandId} and \code{sequenceNo}.

For example: \verb{\\\{strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:14\\\}}}

\item{DigestTipAddress}{The latest block location covered by the digest for which to request a
proof. An address is an Amazon Ion structure that has two fields:
\code{strandId} and \code{sequenceNo}.

For example: \verb{\\\{strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:49\\\}}}
}
\description{
Returns a block object at a specified address in a journal. Also returns
a proof of the specified block for verification if \code{DigestTipAddress} is
provided.
}
\details{
For information about the data contents in a block, see \href{https://docs.aws.amazon.com/qldb/latest/developerguide/journal-contents.html}{Journal contents}
in the \emph{Amazon QLDB Developer Guide}.

If the specified ledger doesn\'t exist or is in \code{DELETING} status, then
throws \code{ResourceNotFoundException}.

If the specified ledger is in \code{CREATING} status, then throws
\code{ResourcePreconditionNotMetException}.

If no block exists with the specified address, then throws
\code{InvalidParameterException}.
}
\section{Request syntax}{
\preformatted{svc$get_block(
  Name = "string",
  BlockAddress = list(
    IonText = "string"
  ),
  DigestTipAddress = list(
    IonText = "string"
  )
)
}
}

\keyword{internal}
