% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_list_journal_kinesis_streams_for_ledger}
\alias{qldb_list_journal_kinesis_streams_for_ledger}
\title{Returns an array of all Amazon QLDB journal stream descriptors for a
given ledger}
\usage{
qldb_list_journal_kinesis_streams_for_ledger(LedgerName, MaxResults,
  NextToken)
}
\arguments{
\item{LedgerName}{[required] The name of the ledger.}

\item{MaxResults}{The maximum number of results to return in a single
\code{ListJournalKinesisStreamsForLedger} request. (The actual number of
results returned might be fewer.)}

\item{NextToken}{A pagination token, indicating that you want to retrieve the next page
of results. If you received a value for \code{NextToken} in the response from
a previous \code{ListJournalKinesisStreamsForLedger} call, you should use
that value as input here.}
}
\description{
Returns an array of all Amazon QLDB journal stream descriptors for a
given ledger. The output of each stream descriptor includes the same
details that are returned by \code{DescribeJournalKinesisStream}.

This action returns a maximum of \code{MaxResults} items. It is paginated so
that you can retrieve all the items by calling
\code{ListJournalKinesisStreamsForLedger} multiple times.
}
\section{Request syntax}{
\preformatted{svc$list_journal_kinesis_streams_for_ledger(
  LedgerName = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
