% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_list_journal_s3_exports_for_ledger}
\alias{qldb_list_journal_s3_exports_for_ledger}
\title{Returns an array of journal export job descriptions for a specified
ledger}
\usage{
qldb_list_journal_s3_exports_for_ledger(Name, MaxResults, NextToken)
}
\arguments{
\item{Name}{[required] The name of the ledger.}

\item{MaxResults}{The maximum number of results to return in a single
\code{\link[=qldb_list_journal_s3_exports_for_ledger]{list_journal_s3_exports_for_ledger}}
request. (The actual number of results returned might be fewer.)}

\item{NextToken}{A pagination token, indicating that you want to retrieve the next page
of results. If you received a value for \code{NextToken} in the response from
a previous
\code{\link[=qldb_list_journal_s3_exports_for_ledger]{list_journal_s3_exports_for_ledger}}
call, then you should use that value as input here.}
}
\description{
Returns an array of journal export job descriptions for a specified
ledger.

This action returns a maximum of \code{MaxResults} items, and is paginated so
that you can retrieve all the items by calling
\code{\link[=qldb_list_journal_s3_exports_for_ledger]{list_journal_s3_exports_for_ledger}}
multiple times.

This action does not return any expired export jobs. For more
information, see \href{https://docs.aws.amazon.com/qldb/latest/developerguide/export-journal.request.html#export-journal.request.expiration}{Export Job Expiration}
in the \emph{Amazon QLDB Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$list_journal_s3_exports_for_ledger(
  Name = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
