% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_list_ledgers}
\alias{qldb_list_ledgers}
\title{Returns an array of ledger summaries that are associated with the
current AWS account and Region}
\usage{
qldb_list_ledgers(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in a single \code{ListLedgers}
request. (The actual number of results returned might be fewer.)}

\item{NextToken}{A pagination token, indicating that you want to retrieve the next page
of results. If you received a value for \code{NextToken} in the response from
a previous \code{ListLedgers} call, then you should use that value as input
here.}
}
\description{
Returns an array of ledger summaries that are associated with the
current AWS account and Region.
}
\details{
This action returns a maximum of 100 items and is paginated so that you
can retrieve all the items by calling \code{ListLedgers} multiple times.
}
\section{Request syntax}{
\preformatted{svc$list_ledgers(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
