% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_tag_resource}
\alias{qldb_tag_resource}
\title{Adds one or more tags to a specified Amazon QLDB resource}
\usage{
qldb_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) to which you want to add the tags. For
example:

\code{arn:aws:qldb:us-east-1:123456789012:ledger/exampleLedger}}

\item{Tags}{[required] The key-value pairs to add as tags to the specified QLDB resource. Tag
keys are case sensitive. If you specify a key that already exists for
the resource, your request fails and returns an error. Tag values are
case sensitive and can be null.}
}
\description{
Adds one or more tags to a specified Amazon QLDB resource.

A resource can have up to 50 tags. If you try to create more than 50
tags for a resource, your request fails and returns an error.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
