% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_update_ledger}
\alias{qldb_update_ledger}
\title{Updates properties on a ledger}
\usage{
qldb_update_ledger(Name, DeletionProtection)
}
\arguments{
\item{Name}{[required] The name of the ledger.}

\item{DeletionProtection}{The flag that prevents a ledger from being deleted by any user. If not
provided on ledger creation, this feature is enabled (\code{true}) by
default.

If deletion protection is enabled, you must first disable it before you
can delete the ledger using the QLDB API or the AWS Command Line
Interface (AWS CLI). You can disable it by calling the
\code{\link[=qldb_update_ledger]{update_ledger}} operation to set the flag to
\code{false}. The QLDB console disables deletion protection for you when you
use it to delete a ledger.}
}
\description{
Updates properties on a ledger.
}
\section{Request syntax}{
\preformatted{svc$update_ledger(
  Name = "string",
  DeletionProtection = TRUE|FALSE
)
}
}

\keyword{internal}
