% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldbsession_service.R
\name{qldbsession}
\alias{qldbsession}
\title{Amazon QLDB Session}
\usage{
qldbsession(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
The transactional data APIs for Amazon QLDB
}
\section{Service syntax}{
\preformatted{svc <- qldbsession(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=qldbsession_send_command]{send_command} \tab Sends a command to an Amazon QLDB ledger
}
}

\examples{
\dontrun{
svc <- qldbsession()
svc$send_command(
  Foo = 123
)
}

}
