% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_iam_policy_assignment}
\alias{quicksight_create_iam_policy_assignment}
\title{Creates an assignment with one specified IAM policy Amazon Resource Name
(ARN) and will assigned to specified groups or users of QuickSight}
\usage{
quicksight_create_iam_policy_assignment(AwsAccountId, AssignmentName,
  AssignmentStatus, PolicyArn, Identities, Namespace)
}
\arguments{
\item{AwsAccountId}{[required] The AWS Account ID where you want to assign QuickSight users or groups
to an IAM policy.}

\item{AssignmentName}{[required] The name of the assignment. It must be unique within an AWS account.}

\item{AssignmentStatus}{[required] The status of an assignment:
\itemize{
\item ENABLED - Anything specified in this assignment is used while
creating the data source.
\item DISABLED - This assignment isn\'t used while creating the data
source.
\item DRAFT - Assignment is an unfinished draft and isn\'t used while
creating the data source.
}}

\item{PolicyArn}{An IAM policy Amazon Resource Name (ARN) that you want to apply to the
QuickSight users and groups specified in this assignment.}

\item{Identities}{QuickSight users and/or groups that you want to assign the policy to.}

\item{Namespace}{[required] The namespace that contains the assignment.}
}
\description{
Creates an assignment with one specified IAM policy Amazon Resource Name
(ARN) and will assigned to specified groups or users of QuickSight.
Users and groups need to be in the same namespace.
}
\section{Request syntax}{
\preformatted{svc$create_iam_policy_assignment(
  AwsAccountId = "string",
  AssignmentName = "string",
  AssignmentStatus = "ENABLED"|"DRAFT"|"DISABLED",
  PolicyArn = "string",
  Identities = list(
    list(
      "string"
    )
  ),
  Namespace = "string"
)
}
}

\keyword{internal}
