% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_account_customization}
\alias{quicksight_delete_account_customization}
\title{Deletes all Amazon QuickSight customizations in this AWS Region for the
specified AWS account and QuickSight namespace}
\usage{
quicksight_delete_account_customization(AwsAccountId, Namespace)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the AWS account that you want to delete QuickSight
customizations from in this AWS Region.}

\item{Namespace}{The QuickSight namespace that you're deleting the customizations from.}
}
\description{
Deletes all Amazon QuickSight customizations in this AWS Region for the
specified AWS account and QuickSight namespace.
}
\section{Request syntax}{
\preformatted{svc$delete_account_customization(
  AwsAccountId = "string",
  Namespace = "string"
)
}
}

\keyword{internal}
