% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_analysis}
\alias{quicksight_describe_analysis}
\title{Provides a summary of the metadata for an analysis}
\usage{
quicksight_describe_analysis(AwsAccountId, AnalysisId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the analysis. You must be using
the AWS account that the analysis is in.}

\item{AnalysisId}{[required] The ID of the analysis that you're describing. The ID is part of the URL
of the analysis.}
}
\description{
Provides a summary of the metadata for an analysis.
}
\section{Request syntax}{
\preformatted{svc$describe_analysis(
  AwsAccountId = "string",
  AnalysisId = "string"
)
}
}

\keyword{internal}
