% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_analysis_permissions}
\alias{quicksight_describe_analysis_permissions}
\title{Provides the read and write permissions for an analysis}
\usage{
quicksight_describe_analysis_permissions(AwsAccountId, AnalysisId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the analysis whose permissions
you're describing. You must be using the AWS account that the analysis
is in.}

\item{AnalysisId}{[required] The ID of the analysis whose permissions you're describing. The ID is
part of the analysis URL.}
}
\description{
Provides the read and write permissions for an analysis.
}
\section{Request syntax}{
\preformatted{svc$describe_analysis_permissions(
  AwsAccountId = "string",
  AnalysisId = "string"
)
}
}

\keyword{internal}
