% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_data_set}
\alias{quicksight_describe_data_set}
\title{Describes a dataset}
\usage{
quicksight_describe_data_set(AwsAccountId, DataSetId)
}
\arguments{
\item{AwsAccountId}{[required] The AWS account ID.}

\item{DataSetId}{[required] The ID for the dataset that you want to create. This ID is unique per
AWS Region for each AWS account.}
}
\description{
Describes a dataset.
}
\section{Request syntax}{
\preformatted{svc$describe_data_set(
  AwsAccountId = "string",
  DataSetId = "string"
)
}
}

\keyword{internal}
