% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_group}
\alias{quicksight_describe_group}
\title{Returns an Amazon QuickSight group's description and Amazon Resource
Name (ARN)}
\usage{
quicksight_describe_group(GroupName, AwsAccountId, Namespace)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to describe.}

\item{AwsAccountId}{[required] The ID for the AWS account that the group is in. Currently, you use the
ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}
}
\description{
Returns an Amazon QuickSight group's description and Amazon Resource
Name (ARN).
}
\details{
The permissions resource is
\code{arn:aws:quicksight:us-east-1:<i>&lt;relevant-aws-account-id&gt;</i>:group/default/<i>&lt;group-name&gt;</i> }.

The response is the group object.

\strong{CLI Sample:}

\code{aws quicksight describe-group -\\-aws-account-id=11112222333 -\\-namespace=default -\\-group-name=Sales }
}
\section{Request syntax}{
\preformatted{svc$describe_group(
  GroupName = "string",
  AwsAccountId = "string",
  Namespace = "string"
)
}
}

\keyword{internal}
