% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_template}
\alias{quicksight_describe_template}
\title{Describes a template's metadata}
\usage{
quicksight_describe_template(AwsAccountId, TemplateId, VersionNumber,
  AliasName)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the template that you\'re
describing.}

\item{TemplateId}{[required] The ID for the template.}

\item{VersionNumber}{(Optional) The number for the version to describe. If a \code{VersionNumber}
parameter value isn\'t provided, the latest version of the template is
described.}

\item{AliasName}{The alias of the template that you want to describe. If you name a
specific alias, you describe the version that the alias points to. You
can specify the latest version of the template by providing the keyword
\verb{$LATEST} in the \code{AliasName} parameter. The keyword \verb{$PUBLISHED}
doesn\'t apply to templates.}
}
\description{
Describes a template\'s metadata.
}
\section{Request syntax}{
\preformatted{svc$describe_template(
  AwsAccountId = "string",
  TemplateId = "string",
  VersionNumber = 123,
  AliasName = "string"
)
}
}

\keyword{internal}
