% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_analyses}
\alias{quicksight_list_analyses}
\title{Lists Amazon QuickSight analyses that exist in the specified AWS account}
\usage{
quicksight_list_analyses(AwsAccountId, NextToken, MaxResults)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the analyses.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Lists Amazon QuickSight analyses that exist in the specified AWS
account.
}
\section{Request syntax}{
\preformatted{svc$list_analyses(
  AwsAccountId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
