% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_data_sources}
\alias{quicksight_list_data_sources}
\title{Lists data sources in current AWS Region that belong to this AWS account}
\usage{
quicksight_list_data_sources(AwsAccountId, NextToken, MaxResults)
}
\arguments{
\item{AwsAccountId}{[required] The AWS account ID.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists data sources in current AWS Region that belong to this AWS
account.
}
\section{Request syntax}{
\preformatted{svc$list_data_sources(
  AwsAccountId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
