% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_template_versions}
\alias{quicksight_list_template_versions}
\title{Lists all the versions of the templates in the Quicksight account}
\usage{
quicksight_list_template_versions(AwsAccountId, TemplateId, NextToken,
  MaxResults)
}
\arguments{
\item{AwsAccountId}{[required] AWS account ID that contains the templates you are listing.}

\item{TemplateId}{[required] The ID for the template.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all the versions of the templates in the Quicksight account.
}
\section{Request syntax}{
\preformatted{svc$list_template_versions(
  AwsAccountId = "string",
  TemplateId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
