% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_theme_aliases}
\alias{quicksight_list_theme_aliases}
\title{Lists all the aliases of a theme}
\usage{
quicksight_list_theme_aliases(AwsAccountId, ThemeId, NextToken,
  MaxResults)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the theme aliases that you\'re
listing.}

\item{ThemeId}{[required] The ID for the theme.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all the aliases of a theme.
}
\section{Request syntax}{
\preformatted{svc$list_theme_aliases(
  AwsAccountId = "string",
  ThemeId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
