% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_delete_resource_share}
\alias{ram_delete_resource_share}
\title{Deletes the specified resource share}
\usage{
ram_delete_resource_share(resourceShareArn, clientToken)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Deletes the specified resource share.
}
\section{Request syntax}{
\preformatted{ram$delete_resource_share(
  resourceShareArn = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
