% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_update_resource_share}
\alias{ram_update_resource_share}
\title{Updates the specified resource share}
\usage{
ram_update_resource_share(resourceShareArn, name,
  allowExternalPrincipals, clientToken)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{name}{The name of the resource share.}

\item{allowExternalPrincipals}{Indicates whether principals outside your organization can be associated
with a resource share.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Updates the specified resource share.
}
\section{Request syntax}{
\preformatted{svc$update_resource_share(
  resourceShareArn = "string",
  name = "string",
  allowExternalPrincipals = TRUE|FALSE,
  clientToken = "string"
)
}
}

\keyword{internal}
