% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_custom_availability_zone}
\alias{rds_delete_custom_availability_zone}
\title{Deletes a custom Availability Zone (AZ)}
\usage{
rds_delete_custom_availability_zone(CustomAvailabilityZoneId)
}
\arguments{
\item{CustomAvailabilityZoneId}{[required] The custom AZ identifier.}
}
\description{
Deletes a custom Availability Zone (AZ).

A custom AZ is an on-premises AZ that is integrated with a VMware
vSphere cluster.

For more information about RDS on VMware, see the \href{https://docs.aws.amazon.com/AmazonRDS/latest/RDSonVMwareUserGuide/rds-on-vmware.html}{RDS on VMware User Guide.}
}
\section{Request syntax}{
\preformatted{svc$delete_custom_availability_zone(
  CustomAvailabilityZoneId = "string"
)
}
}

\keyword{internal}
