% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_cluster_endpoint}
\alias{rds_delete_db_cluster_endpoint}
\title{Deletes a custom endpoint and removes it from an Amazon Aurora DB
cluster}
\usage{
rds_delete_db_cluster_endpoint(DBClusterEndpointIdentifier)
}
\arguments{
\item{DBClusterEndpointIdentifier}{[required] The identifier associated with the custom endpoint. This parameter is
stored as a lowercase string.}
}
\description{
Deletes a custom endpoint and removes it from an Amazon Aurora DB
cluster.
}
\details{
This action only applies to Aurora DB clusters.
}
\section{Request syntax}{
\preformatted{svc$delete_db_cluster_endpoint(
  DBClusterEndpointIdentifier = "string"
)
}
}

\keyword{internal}
