% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_modify_db_instance}
\alias{rds_modify_db_instance}
\title{Modifies settings for a DB instance}
\usage{
rds_modify_db_instance(DBInstanceIdentifier, AllocatedStorage,
  DBInstanceClass, DBSubnetGroupName, DBSecurityGroups,
  VpcSecurityGroupIds, ApplyImmediately, MasterUserPassword,
  DBParameterGroupName, BackupRetentionPeriod, PreferredBackupWindow,
  PreferredMaintenanceWindow, MultiAZ, EngineVersion,
  AllowMajorVersionUpgrade, AutoMinorVersionUpgrade, LicenseModel, Iops,
  OptionGroupName, NewDBInstanceIdentifier, StorageType, TdeCredentialArn,
  TdeCredentialPassword, CACertificateIdentifier, Domain,
  CopyTagsToSnapshot, MonitoringInterval, DBPortNumber,
  PubliclyAccessible, MonitoringRoleArn, DomainIAMRoleName, PromotionTier,
  EnableIAMDatabaseAuthentication, EnablePerformanceInsights,
  PerformanceInsightsKMSKeyId, PerformanceInsightsRetentionPeriod,
  CloudwatchLogsExportConfiguration, ProcessorFeatures,
  UseDefaultProcessorFeatures, DeletionProtection)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The DB instance identifier. This value is stored as a lowercase string.

Constraints:
\itemize{
\item Must match the identifier of an existing DBInstance.
}}

\item{AllocatedStorage}{The new amount of storage (in gibibytes) to allocate for the DB
instance.

For MariaDB, MySQL, Oracle, and PostgreSQL, the value supplied must be
at least 10\% greater than the current value. Values that are not at
least 10\% greater than the existing value are rounded up so that they
are 10\% greater than the current value.

For the valid values for allocated storage for each engine, see
CreateDBInstance.}

\item{DBInstanceClass}{The new compute and memory capacity of the DB instance, for example,
\code{db.m4.large}. Not all DB instance classes are available in all AWS
Regions, or for all database engines. For the full list of DB instance
classes, and availability for your engine, see \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/Concepts.DBInstanceClass.html}{DB Instance Class}
in the \emph{Amazon RDS User Guide.}

If you modify the DB instance class, an outage occurs during the change.
The change is applied during the next maintenance window, unless
\code{ApplyImmediately} is specified as \code{true} for this request.

Default: Uses existing setting}

\item{DBSubnetGroupName}{The new DB subnet group for the DB instance. You can use this parameter
to move your DB instance to a different VPC. If your DB instance is not
in a VPC, you can also use this parameter to move your DB instance into
a VPC. For more information, see \href{http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_VPC.WorkingWithRDSInstanceinaVPC.html#USER_VPC.Non-VPC2VPC}{Updating the VPC for a DB Instance}
in the \emph{Amazon RDS User Guide.}

Changing the subnet group causes an outage during the change. The change
is applied during the next maintenance window, unless you specify \code{true}
for the \code{ApplyImmediately} parameter.

Constraints: If supplied, must match the name of an existing
DBSubnetGroup.

Example: \code{mySubnetGroup}}

\item{DBSecurityGroups}{A list of DB security groups to authorize on this DB instance. Changing
this setting doesn't result in an outage and the change is
asynchronously applied as soon as possible.

Constraints:
\itemize{
\item If supplied, must match existing DBSecurityGroups.
}}

\item{VpcSecurityGroupIds}{A list of EC2 VPC security groups to authorize on this DB instance. This
change is asynchronously applied as soon as possible.

\strong{Amazon Aurora}

Not applicable. The associated list of EC2 VPC security groups is
managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:
\itemize{
\item If supplied, must match existing VpcSecurityGroupIds.
}}

\item{ApplyImmediately}{Specifies whether the modifications in this request and any pending
modifications are asynchronously applied as soon as possible, regardless
of the \code{PreferredMaintenanceWindow} setting for the DB instance.

If this parameter is set to \code{false}, changes to the DB instance are
applied during the next maintenance window. Some parameter changes can
cause an outage and are applied on the next call to RebootDBInstance, or
the next failure reboot. Review the table of parameters in \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/Overview.DBInstance.Modifying.html}{Modifying a DB Instance and Using the Apply Immediately Parameter}
in the \emph{Amazon RDS User Guide.} to see the impact that setting
\code{ApplyImmediately} to \code{true} or \code{false} has for each modified parameter
and to determine when the changes are applied.

Default: \code{false}}

\item{MasterUserPassword}{The new password for the master user. The password can include any
printable ASCII character except "/", """, or "@".

Changing this parameter doesn't result in an outage and the change is
asynchronously applied as soon as possible. Between the time of the
request and the completion of the request, the \code{MasterUserPassword}
element exists in the \code{PendingModifiedValues} element of the operation
response.

\strong{Amazon Aurora}

Not applicable. The password for the master user is managed by the DB
cluster. For more information, see ModifyDBCluster.

Default: Uses existing setting

\strong{MariaDB}

Constraints: Must contain from 8 to 41 characters.

\strong{Microsoft SQL Server}

Constraints: Must contain from 8 to 128 characters.

\strong{MySQL}

Constraints: Must contain from 8 to 41 characters.

\strong{Oracle}

Constraints: Must contain from 8 to 30 characters.

\strong{PostgreSQL}

Constraints: Must contain from 8 to 128 characters.

Amazon RDS API actions never return the password, so this action
provides a way to regain access to a primary instance user if the
password is lost. This includes restoring privileges that might have
been accidentally revoked.}

\item{DBParameterGroupName}{The name of the DB parameter group to apply to the DB instance. Changing
this setting doesn't result in an outage. The parameter group name
itself is changed immediately, but the actual parameter changes are not
applied until you reboot the instance without failover. The db instance
will NOT be rebooted automatically and the parameter changes will NOT be
applied during the next maintenance window.

Default: Uses existing setting

Constraints: The DB parameter group must be in the same DB parameter
group family as this DB instance.}

\item{BackupRetentionPeriod}{The number of days to retain automated backups. Setting this parameter
to a positive number enables backups. Setting this parameter to 0
disables automated backups.

Changing this parameter can result in an outage if you change from 0 to
a non-zero value or from a non-zero value to 0. These changes are
applied during the next maintenance window unless the \code{ApplyImmediately}
parameter is set to \code{true} for this request. If you change the parameter
from one non-zero value to another non-zero value, the change is
asynchronously applied as soon as possible.

\strong{Amazon Aurora}

Not applicable. The retention period for automated backups is managed by
the DB cluster. For more information, see ModifyDBCluster.

Default: Uses existing setting

Constraints:
\itemize{
\item Must be a value from 0 to 35
\item Can be specified for a MySQL Read Replica only if the source is
running MySQL 5.6 or later
\item Can be specified for a PostgreSQL Read Replica only if the source is
running PostgreSQL 9.3.5
\item Can't be set to 0 if the DB instance is a source to Read Replicas
}}

\item{PreferredBackupWindow}{The daily time range during which automated backups are created if
automated backups are enabled, as determined by the
\code{BackupRetentionPeriod} parameter. Changing this parameter doesn't
result in an outage and the change is asynchronously applied as soon as
possible.

\strong{Amazon Aurora}

Not applicable. The daily time range for creating automated backups is
managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:
\itemize{
\item Must be in the format hh24:mi-hh24:mi
\item Must be in Universal Time Coordinated (UTC)
\item Must not conflict with the preferred maintenance window
\item Must be at least 30 minutes
}}

\item{PreferredMaintenanceWindow}{The weekly time range (in UTC) during which system maintenance can
occur, which might result in an outage. Changing this parameter doesn't
result in an outage, except in the following situation, and the change
is asynchronously applied as soon as possible. If there are pending
actions that cause a reboot, and the maintenance window is changed to
include the current time, then changing this parameter will cause a
reboot of the DB instance. If moving this window to the current time,
there must be at least 30 minutes between the current time and end of
the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon \| Tue \| Wed \| Thu \| Fri \| Sat \| Sun

Constraints: Must be at least 30 minutes}

\item{MultiAZ}{Specifies if the DB instance is a Multi-AZ deployment. Changing this
parameter doesn't result in an outage and the change is applied during
the next maintenance window unless the \code{ApplyImmediately} parameter is
set to \code{true} for this request.}

\item{EngineVersion}{The version number of the database engine to upgrade to. Changing this
parameter results in an outage and the change is applied during the next
maintenance window unless the \code{ApplyImmediately} parameter is set to
\code{true} for this request.

For major version upgrades, if a nondefault DB parameter group is
currently in use, a new DB parameter group in the DB parameter group
family for the new engine version must be specified. The new DB
parameter group can be the default for that DB parameter group family.

For information about valid engine versions, see CreateDBInstance, or
call DescribeDBEngineVersions.}

\item{AllowMajorVersionUpgrade}{Indicates that major version upgrades are allowed. Changing this
parameter doesn't result in an outage and the change is asynchronously
applied as soon as possible.

Constraints: This parameter must be set to true when specifying a value
for the EngineVersion parameter that is a different major version than
the DB instance's current version.}

\item{AutoMinorVersionUpgrade}{Indicates that minor version upgrades are applied automatically to the
DB instance during the maintenance window. Changing this parameter
doesn't result in an outage except in the following case and the change
is asynchronously applied as soon as possible. An outage will result if
this parameter is set to \code{true} during the maintenance window, and a
newer minor version is available, and RDS has enabled auto patching for
that engine version.}

\item{LicenseModel}{The license model for the DB instance.

Valid values: \code{license-included} \| \code{bring-your-own-license} \|
\code{general-public-license}}

\item{Iops}{The new Provisioned IOPS (I/O operations per second) value for the RDS
instance.

Changing this setting doesn't result in an outage and the change is
applied during the next maintenance window unless the \code{ApplyImmediately}
parameter is set to \code{true} for this request. If you are migrating from
Provisioned IOPS to standard storage, set this value to 0. The DB
instance will require a reboot for the change in storage type to take
effect.

If you choose to migrate your DB instance from using standard storage to
using Provisioned IOPS, or from using Provisioned IOPS to using standard
storage, the process can take time. The duration of the migration
depends on several factors such as database load, storage size, storage
type (standard or Provisioned IOPS), amount of IOPS provisioned (if
any), and the number of prior scale storage operations. Typical
migration times are under 24 hours, but the process can take up to
several days in some cases. During the migration, the DB instance is
available for use, but might experience performance degradation. While
the migration takes place, nightly backups for the instance are
suspended. No other Amazon RDS operations can take place for the
instance, including modifying the instance, rebooting the instance,
deleting the instance, creating a Read Replica for the instance, and
creating a DB snapshot of the instance.

Constraints: For MariaDB, MySQL, Oracle, and PostgreSQL, the value
supplied must be at least 10\% greater than the current value. Values
that are not at least 10\% greater than the existing value are rounded up
so that they are 10\% greater than the current value.

Default: Uses existing setting}

\item{OptionGroupName}{Indicates that the DB instance should be associated with the specified
option group. Changing this parameter doesn't result in an outage
except in the following case and the change is applied during the next
maintenance window unless the \code{ApplyImmediately} parameter is set to
\code{true} for this request. If the parameter change results in an option
group that enables OEM, this change can cause a brief (sub-second)
period during which new connections are rejected but existing
connections are not interrupted.

Permanent options, such as the TDE option for Oracle Advanced Security
TDE, can't be removed from an option group, and that option group
can't be removed from a DB instance once it is associated with a DB
instance}

\item{NewDBInstanceIdentifier}{The new DB instance identifier for the DB instance when renaming a DB
instance. When you change the DB instance identifier, an instance reboot
will occur immediately if you set \code{Apply Immediately} to true, or will
occur during the next maintenance window if \code{Apply Immediately} to
false. This value is stored as a lowercase string.

Constraints:
\itemize{
\item Must contain from 1 to 63 letters, numbers, or hyphens.
\item The first character must be a letter.
\item Can't end with a hyphen or contain two consecutive hyphens.
}

Example: \code{mydbinstance}}

\item{StorageType}{Specifies the storage type to be associated with the DB instance.

If you specify Provisioned IOPS (\code{io1}), you must also include a value
for the \code{Iops} parameter.

If you choose to migrate your DB instance from using standard storage to
using Provisioned IOPS, or from using Provisioned IOPS to using standard
storage, the process can take time. The duration of the migration
depends on several factors such as database load, storage size, storage
type (standard or Provisioned IOPS), amount of IOPS provisioned (if
any), and the number of prior scale storage operations. Typical
migration times are under 24 hours, but the process can take up to
several days in some cases. During the migration, the DB instance is
available for use, but might experience performance degradation. While
the migration takes place, nightly backups for the instance are
suspended. No other Amazon RDS operations can take place for the
instance, including modifying the instance, rebooting the instance,
deleting the instance, creating a Read Replica for the instance, and
creating a DB snapshot of the instance.

Valid values: \code{standard | gp2 | io1}

Default: \code{io1} if the \code{Iops} parameter is specified, otherwise
\code{standard}}

\item{TdeCredentialArn}{The ARN from the key store with which to associate the instance for TDE
encryption.}

\item{TdeCredentialPassword}{The password for the given ARN from the key store in order to access the
device.}

\item{CACertificateIdentifier}{Indicates the certificate that needs to be associated with the instance.}

\item{Domain}{The Active Directory Domain to move the instance to. Specify \code{none} to
remove the instance from its current domain. The domain must be created
prior to this operation. Currently only a Microsoft SQL Server instance
can be created in a Active Directory Domain.}

\item{CopyTagsToSnapshot}{True to copy all tags from the DB instance to snapshots of the DB
instance, and otherwise false. The default is false.

\strong{Amazon Aurora}

Not applicable. Copying tags to snapshots is managed by the DB cluster.
Setting this value for an Aurora DB instance has no effect on the DB
cluster setting. For more information, see ModifyDBCluster.}

\item{MonitoringInterval}{The interval, in seconds, between points when Enhanced Monitoring
metrics are collected for the DB instance. To disable collecting
Enhanced Monitoring metrics, specify 0. The default is 0.

If \code{MonitoringRoleArn} is specified, then you must also set
\code{MonitoringInterval} to a value other than 0.

Valid Values: \code{0, 1, 5, 10, 15, 30, 60}}

\item{DBPortNumber}{The port number on which the database accepts connections.

The value of the \code{DBPortNumber} parameter must not match any of the port
values specified for options in the option group for the DB instance.

Your database will restart when you change the \code{DBPortNumber} value
regardless of the value of the \code{ApplyImmediately} parameter.

\strong{MySQL}

Default: \code{3306}

Valid Values: \code{1150-65535}

\strong{MariaDB}

Default: \code{3306}

Valid Values: \code{1150-65535}

\strong{PostgreSQL}

Default: \code{5432}

Valid Values: \code{1150-65535}

Type: Integer

\strong{Oracle}

Default: \code{1521}

Valid Values: \code{1150-65535}

\strong{SQL Server}

Default: \code{1433}

Valid Values: \code{1150-65535} except for \code{1434}, \code{3389}, \code{47001}, \code{49152},
and \code{49152} through \code{49156}.

\strong{Amazon Aurora}

Default: \code{3306}

Valid Values: \code{1150-65535}}

\item{PubliclyAccessible}{Boolean value that indicates if the DB instance has a publicly
resolvable DNS name. Set to \code{True} to make the DB instance
Internet-facing with a publicly resolvable DNS name, which resolves to a
public IP address. Set to \code{False} to make the DB instance internal with
a DNS name that resolves to a private IP address.

\code{PubliclyAccessible} only applies to DB instances in a VPC. The DB
instance must be part of a public subnet and \code{PubliclyAccessible} must
be true in order for it to be publicly accessible.

Changes to the \code{PubliclyAccessible} parameter are applied immediately
regardless of the value of the \code{ApplyImmediately} parameter.

Default: false}

\item{MonitoringRoleArn}{The ARN for the IAM role that permits RDS to send enhanced monitoring
metrics to Amazon CloudWatch Logs. For example,
\code{arn:aws:iam:123456789012:role/emaccess}. For information on creating a
monitoring role, go to \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_Monitoring.html#USER_Monitoring.OS.IAMRole}{To create an IAM role for Amazon RDS Enhanced Monitoring}
in the \emph{Amazon RDS User Guide.}

If \code{MonitoringInterval} is set to a value other than 0, then you must
supply a \code{MonitoringRoleArn} value.}

\item{DomainIAMRoleName}{The name of the IAM role to use when making API calls to the Directory
Service.}

\item{PromotionTier}{A value that specifies the order in which an Aurora Replica is promoted
to the primary instance after a failure of the existing primary
instance. For more information, see \href{https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/Aurora.Managing.Backups.html#Aurora.Managing.FaultTolerance}{Fault Tolerance for an Aurora DB Cluster}
in the \emph{Amazon Aurora User Guide}.

Default: 1

Valid Values: 0 - 15}

\item{EnableIAMDatabaseAuthentication}{True to enable mapping of AWS Identity and Access Management (IAM)
accounts to database accounts, and otherwise false.

You can enable IAM database authentication for the following database
engines

\strong{Amazon Aurora}

Not applicable. Mapping AWS IAM accounts to database accounts is managed
by the DB cluster. For more information, see ModifyDBCluster.

\strong{MySQL}
\itemize{
\item For MySQL 5.6, minor version 5.6.34 or higher
\item For MySQL 5.7, minor version 5.7.16 or higher
}

Default: \code{false}}

\item{EnablePerformanceInsights}{True to enable Performance Insights for the DB instance, and otherwise
false.

For more information, see \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_PerfInsights.html}{Using Amazon Performance Insights}
in the \emph{Amazon Relational Database Service User Guide}.}

\item{PerformanceInsightsKMSKeyId}{The AWS KMS key identifier for encryption of Performance Insights data.
The KMS key ID is the Amazon Resource Name (ARN), KMS key identifier, or
the KMS key alias for the KMS encryption key.}

\item{PerformanceInsightsRetentionPeriod}{The amount of time, in days, to retain Performance Insights data. Valid
values are 7 or 731 (2 years).}

\item{CloudwatchLogsExportConfiguration}{The configuration setting for the log types to be enabled for export to
CloudWatch Logs for a specific DB instance.

A change to the \code{CloudwatchLogsExportConfiguration} parameter is always
applied to the DB instance immediately. Therefore, the
\code{ApplyImmediately} parameter has no effect.}

\item{ProcessorFeatures}{The number of CPU cores and the number of threads per core for the DB
instance class of the DB instance.}

\item{UseDefaultProcessorFeatures}{A value that specifies that the DB instance class of the DB instance
uses its default processor features.}

\item{DeletionProtection}{Indicates if the DB instance has deletion protection enabled. The
database can't be deleted when this value is set to true. For more
information, see \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_DeleteInstance.html}{Deleting a DB Instance}.}
}
\description{
Modifies settings for a DB instance. You can change one or more database
configuration parameters by specifying these parameters and the new
values in the request. To learn what modifications you can make to your
DB instance, call DescribeValidDBInstanceModifications before you call
ModifyDBInstance.
}
\section{Request syntax}{
\preformatted{svc$modify_db_instance(
  DBInstanceIdentifier = "string",
  AllocatedStorage = 123,
  DBInstanceClass = "string",
  DBSubnetGroupName = "string",
  DBSecurityGroups = list(
    "string"
  ),
  VpcSecurityGroupIds = list(
    "string"
  ),
  ApplyImmediately = TRUE|FALSE,
  MasterUserPassword = "string",
  DBParameterGroupName = "string",
  BackupRetentionPeriod = 123,
  PreferredBackupWindow = "string",
  PreferredMaintenanceWindow = "string",
  MultiAZ = TRUE|FALSE,
  EngineVersion = "string",
  AllowMajorVersionUpgrade = TRUE|FALSE,
  AutoMinorVersionUpgrade = TRUE|FALSE,
  LicenseModel = "string",
  Iops = 123,
  OptionGroupName = "string",
  NewDBInstanceIdentifier = "string",
  StorageType = "string",
  TdeCredentialArn = "string",
  TdeCredentialPassword = "string",
  CACertificateIdentifier = "string",
  Domain = "string",
  CopyTagsToSnapshot = TRUE|FALSE,
  MonitoringInterval = 123,
  DBPortNumber = 123,
  PubliclyAccessible = TRUE|FALSE,
  MonitoringRoleArn = "string",
  DomainIAMRoleName = "string",
  PromotionTier = 123,
  EnableIAMDatabaseAuthentication = TRUE|FALSE,
  EnablePerformanceInsights = TRUE|FALSE,
  PerformanceInsightsKMSKeyId = "string",
  PerformanceInsightsRetentionPeriod = 123,
  CloudwatchLogsExportConfiguration = list(
    EnableLogTypes = list(
      "string"
    ),
    DisableLogTypes = list(
      "string"
    )
  ),
  ProcessorFeatures = list(
    list(
      Name = "string",
      Value = "string"
    )
  ),
  UseDefaultProcessorFeatures = TRUE|FALSE,
  DeletionProtection = TRUE|FALSE
)
}
}

\keyword{internal}
