% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsdataservice_operations.R
\name{rdsdataservice_execute_sql}
\alias{rdsdataservice_execute_sql}
\title{Runs one or more SQL statements}
\usage{
rdsdataservice_execute_sql(awsSecretStoreArn, database,
  dbClusterOrInstanceArn, schema, sqlStatements)
}
\arguments{
\item{awsSecretStoreArn}{[required] The Amazon Resource Name (ARN) of the secret that enables access to the
DB cluster.}

\item{database}{The name of the database.}

\item{dbClusterOrInstanceArn}{[required] The ARN of the Aurora Serverless DB cluster.}

\item{schema}{The name of the database schema.}

\item{sqlStatements}{[required] One or more SQL statements to run on the DB cluster.

You can separate SQL statements from each other with a semicolon (;).
Any valid SQL statement is permitted, including data definition, data
manipulation, and commit statements.}
}
\description{
Runs one or more SQL statements.
}
\details{
This operation is deprecated. Use the \code{BatchExecuteStatement} or
\code{ExecuteStatement} operation.
}
\section{Request syntax}{
\preformatted{svc$execute_sql(
  awsSecretStoreArn = "string",
  database = "string",
  dbClusterOrInstanceArn = "string",
  schema = "string",
  sqlStatements = "string"
)
}
}

\keyword{internal}
