% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsdataservice_operations.R
\name{rdsdataservice_execute_statement}
\alias{rdsdataservice_execute_statement}
\title{Runs a SQL statement against a database}
\usage{
rdsdataservice_execute_statement(continueAfterTimeout, database,
  includeResultMetadata, parameters, resourceArn, schema, secretArn, sql,
  transactionId)
}
\arguments{
\item{continueAfterTimeout}{A value that indicates whether to continue running the statement after
the call times out. By default, the statement stops running when the
call times out.

For DDL statements, we recommend continuing to run the statement after
the call times out. When a DDL statement terminates before it is
finished running, it can result in errors and possibly corrupted data
structures.}

\item{database}{The name of the database.}

\item{includeResultMetadata}{A value that indicates whether to include metadata in the results.}

\item{parameters}{The parameters for the SQL statement.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.}

\item{schema}{The name of the database schema.}

\item{secretArn}{[required] The name or ARN of the secret that enables access to the DB cluster.}

\item{sql}{[required] The SQL statement to run.}

\item{transactionId}{The identifier of a transaction that was started by using the
\code{BeginTransaction} operation. Specify the transaction ID of the
transaction that you want to include the SQL statement in.

If the SQL statement is not part of a transaction, don't set this
parameter.}
}
\description{
Runs a SQL statement against a database.
}
\details{
If a call isn't part of a transaction because it doesn't include the
\code{transactionID} parameter, changes that result from the call are
committed automatically.

The response size limit is 1 MB or 1,000 records. If the call returns
more than 1 MB of response data or over 1,000 records, the call is
terminated.
}
\section{Request syntax}{
\preformatted{svc$execute_statement(
  continueAfterTimeout = TRUE|FALSE,
  database = "string",
  includeResultMetadata = TRUE|FALSE,
  parameters = list(
    list(
      name = "string",
      value = list(
        blobValue = raw,
        booleanValue = TRUE|FALSE,
        doubleValue = 123.0,
        isNull = TRUE|FALSE,
        longValue = 123,
        stringValue = "string"
      )
    )
  ),
  resourceArn = "string",
  schema = "string",
  secretArn = "string",
  sql = "string",
  transactionId = "string"
)
}
}

\keyword{internal}
