% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_cluster_parameter_group}
\alias{redshift_delete_cluster_parameter_group}
\title{Deletes a specified Amazon Redshift parameter group}
\usage{
redshift_delete_cluster_parameter_group(ParameterGroupName)
}
\arguments{
\item{ParameterGroupName}{[required] The name of the parameter group to be deleted.

Constraints:
\itemize{
\item Must be the name of an existing cluster parameter group.
\item Cannot delete a default cluster parameter group.
}}
}
\description{
Deletes a specified Amazon Redshift parameter group.

You cannot delete a parameter group if it is associated with a cluster.
}
\section{Request syntax}{
\preformatted{svc$delete_cluster_parameter_group(
  ParameterGroupName = "string"
)
}
}

\keyword{internal}
