% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_cluster_tracks}
\alias{redshift_describe_cluster_tracks}
\title{Returns a list of all the available maintenance tracks}
\usage{
redshift_describe_cluster_tracks(MaintenanceTrackName, MaxRecords,
  Marker)
}
\arguments{
\item{MaintenanceTrackName}{The name of the maintenance track.}

\item{MaxRecords}{An integer value for the maximum number of maintenance tracks to return.}

\item{Marker}{An optional parameter that specifies the starting point to return a set
of response records. When the results of a \code{DescribeClusterTracks}
request exceed the value specified in \code{MaxRecords}, Amazon Redshift
returns a value in the \code{Marker} field of the response. You can retrieve
the next set of response records by providing the returned marker value
in the \code{Marker} parameter and retrying the request.}
}
\description{
Returns a list of all the available maintenance tracks.
}
\section{Request syntax}{
\preformatted{redshift$describe_cluster_tracks(
  MaintenanceTrackName = "string",
  MaxRecords = 123,
  Marker = "string"
)
}
}

\keyword{internal}
