% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_modify_cluster_db_revision}
\alias{redshift_modify_cluster_db_revision}
\title{Modifies the database revision of a cluster}
\usage{
redshift_modify_cluster_db_revision(ClusterIdentifier, RevisionTarget)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier of a cluster whose database revision you want to
modify.

Example: \code{examplecluster}}

\item{RevisionTarget}{[required] The identifier of the database revision. You can retrieve this value
from the response to the DescribeClusterDbRevisions request.}
}
\description{
Modifies the database revision of a cluster. The database revision is a
unique revision of the database running in a cluster.
}
\section{Request syntax}{
\preformatted{redshift$modify_cluster_db_revision(
  ClusterIdentifier = "string",
  RevisionTarget = "string"
)
}
}

\keyword{internal}
