% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_pause_cluster}
\alias{redshift_pause_cluster}
\title{Pauses a cluster}
\usage{
redshift_pause_cluster(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The identifier of the cluster to be paused.}
}
\description{
Pauses a cluster.
}
\section{Request syntax}{
\preformatted{svc$pause_cluster(
  ClusterIdentifier = "string"
)
}
}

\keyword{internal}
