% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftdataapiservice_operations.R
\name{redshiftdataapiservice_describe_statement}
\alias{redshiftdataapiservice_describe_statement}
\title{Describes the details about a specific instance when a query was run by
the Amazon Redshift Data API}
\usage{
redshiftdataapiservice_describe_statement(Id)
}
\arguments{
\item{Id}{[required] The identifier of the SQL statement to describe. This value is a
universally unique identifier (UUID) generated by Amazon Redshift Data
API. This identifier is returned by \code{ExecuteStatment} and
\code{ListStatements}.}
}
\description{
Describes the details about a specific instance when a query was run by
the Amazon Redshift Data API. The information includes when the query
started, when it finished, the query status, the number of rows
returned, and the SQL statement.
}
\section{Request syntax}{
\preformatted{svc$describe_statement(
  Id = "string"
)
}
}

\keyword{internal}
