% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftdataapiservice_operations.R
\name{redshiftdataapiservice_describe_table}
\alias{redshiftdataapiservice_describe_table}
\title{Describes the detailed information about a table from metadata in the
cluster}
\usage{
redshiftdataapiservice_describe_table(ClusterIdentifier, Database,
  DbUser, MaxResults, NextToken, Schema, SecretArn, Table)
}
\arguments{
\item{ClusterIdentifier}{[required] The cluster identifier. This parameter is required when authenticating
using either AWS Secrets Manager or temporary credentials.}

\item{Database}{The name of the database. This parameter is required when authenticating
using temporary credentials.}

\item{DbUser}{The database user name. This parameter is required when authenticating
using temporary credentials.}

\item{MaxResults}{The maximum number of tables to return in the response. If more tables
exist than fit in one response, then \code{NextToken} is returned to page
through the results.}

\item{NextToken}{A value that indicates the starting point for the next set of response
records in a subsequent request. If a value is returned in a response,
you can retrieve the next set of records by providing this returned
NextToken value in the next NextToken parameter and retrying the
command. If the NextToken field is empty, all response records have been
retrieved for the request.}

\item{Schema}{The schema that contains the table. If no schema is specified, then
matching tables for all schemas are returned.}

\item{SecretArn}{The name or ARN of the secret that enables access to the database. This
parameter is required when authenticating using AWS Secrets Manager.}

\item{Table}{The table name. If no table is specified, then all tables for all
matching schemas are returned. If no table and no schema is specified,
then all tables for all schemas in the database are returned}
}
\description{
Describes the detailed information about a table from metadata in the
cluster. The information includes its columns. A token is returned to
page through the column list. Depending on the authorization method, use
one of the following combinations of request parameters:
}
\details{
\itemize{
\item AWS Secrets Manager - specify the Amazon Resource Name (ARN) of the
secret and the cluster identifier that matches the cluster in the
secret.
\item Temporary credentials - specify the cluster identifier, the database
name, and the database user name. Permission to call the
\code{redshift:GetClusterCredentials} operation is required to use this
method.
}
}
\section{Request syntax}{
\preformatted{svc$describe_table(
  ClusterIdentifier = "string",
  Database = "string",
  DbUser = "string",
  MaxResults = 123,
  NextToken = "string",
  Schema = "string",
  SecretArn = "string",
  Table = "string"
)
}
}

\keyword{internal}
