% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftdataapiservice_operations.R
\name{redshiftdataapiservice_get_statement_result}
\alias{redshiftdataapiservice_get_statement_result}
\title{Fetches the temporarily cached result of an SQL statement}
\usage{
redshiftdataapiservice_get_statement_result(Id, NextToken)
}
\arguments{
\item{Id}{[required] The identifier of the SQL statement whose results are to be fetched.
This value is a universally unique identifier (UUID) generated by Amazon
Redshift Data API. This identifier is returned by \code{ExecuteStatment} and
\code{ListStatements}.}

\item{NextToken}{A value that indicates the starting point for the next set of response
records in a subsequent request. If a value is returned in a response,
you can retrieve the next set of records by providing this returned
NextToken value in the next NextToken parameter and retrying the
command. If the NextToken field is empty, all response records have been
retrieved for the request.}
}
\description{
Fetches the temporarily cached result of an SQL statement. A token is
returned to page through the statement results.
}
\section{Request syntax}{
\preformatted{svc$get_statement_result(
  Id = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
