% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_create_project}
\alias{rekognition_create_project}
\title{Creates a new Amazon Rekognition Custom Labels project}
\usage{
rekognition_create_project(ProjectName)
}
\arguments{
\item{ProjectName}{[required] The name of the project to create.}
}
\description{
Creates a new Amazon Rekognition Custom Labels project. A project is a
logical grouping of resources (images, Labels, models) and operations
(training, evaluation and detection).
}
\details{
This operation requires permissions to perform the
\code{rekognition:CreateProject} action.
}
\section{Request syntax}{
\preformatted{svc$create_project(
  ProjectName = "string"
)
}
}

\keyword{internal}
