% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_delete_collection}
\alias{rekognition_delete_collection}
\title{Deletes the specified collection}
\usage{
rekognition_delete_collection(CollectionId)
}
\arguments{
\item{CollectionId}{[required] ID of the collection to delete.}
}
\description{
Deletes the specified collection. Note that this operation removes all
faces in the collection. For an example, see
delete-collection-procedure.
}
\details{
This operation requires permissions to perform the
\code{rekognition:DeleteCollection} action.
}
\section{Request syntax}{
\preformatted{rekognition$delete_collection(
  CollectionId = "string"
)
}
}

\examples{
# This operation deletes a Rekognition collection.
\donttest{rekognition$delete_collection(
  CollectionId = "myphotos"
)}

}
\keyword{internal}
