% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{resourceexplorer}
\alias{resourceexplorer}
\title{AWS Resource Explorer}
\usage{
resourceexplorer(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
\item{\strong{use_dual_stack}: Set this to \code{true} to use the dualstack (IPv4 and IPv6) endpoint for a service, where available, falling back to the regular endpoint when it isn't. Defaults to the \code{AWS_USE_DUALSTACK_ENDPOINT} environment variable when unset.}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Amazon Web Services Resource Explorer is a resource search and discovery service. By using Resource Explorer, you can explore your resources using an internet search engine-like experience. Examples of resources include Amazon Relational Database Service (Amazon RDS) instances, Amazon Simple Storage Service (Amazon S3) buckets, or Amazon DynamoDB tables. You can search for your resources using resource metadata like names, tags, and IDs. Resource Explorer can search across all of the Amazon Web Services Regions in your account in which you turn the service on, to simplify your cross-Region workloads.

Resource Explorer scans the resources in each of the Amazon Web Services Regions in your Amazon Web Services account in which you turn on Resource Explorer. Resource Explorer \href{https://docs.aws.amazon.com/resource-explorer/latest/userguide/getting-started-terms-and-concepts.html#term-index}{creates and maintains an index} in each Region, with the details of that Region's resources.

You can \href{https://docs.aws.amazon.com/resource-explorer/latest/userguide/manage-aggregator-region.html}{search across all of the indexed Regions in your account} by designating one of your Amazon Web Services Regions to contain the aggregator index for the account. When you \href{https://docs.aws.amazon.com/resource-explorer/latest/userguide/manage-aggregator-region-turn-on.html}{promote a local index in a Region to become the aggregator index for the account}, Resource Explorer automatically replicates the index information from all local indexes in the other Regions to the aggregator index. Therefore, the Region with the aggregator index has a copy of all resource information for all Regions in the account where you turned on Resource Explorer. As a result, views in the aggregator index Region include resources from all of the indexed Regions in your account.

For more information about Amazon Web Services Resource Explorer, including how to enable and configure the service, see the \href{https://docs.aws.amazon.com/resource-explorer/latest/userguide/}{Amazon Web Services Resource Explorer User Guide}.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- resourceexplorer(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string",
    use_dual_stack = "logical"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.application.integration:resourceexplorer_associate_default_view]{associate_default_view} \tab Sets the specified view as the default for the Amazon Web Services Region in which you call this operation\cr
\link[paws.application.integration:resourceexplorer_batch_get_view]{batch_get_view} \tab Retrieves details about a list of views\cr
\link[paws.application.integration:resourceexplorer_create_index]{create_index} \tab Turns on Amazon Web Services Resource Explorer in the Amazon Web Services Region in which you called this operation by creating an index\cr
\link[paws.application.integration:resourceexplorer_create_resource_explorer_setup]{create_resource_explorer_setup} \tab Creates a Resource Explorer setup configuration across multiple Amazon Web Services Regions\cr
\link[paws.application.integration:resourceexplorer_create_view]{create_view} \tab Creates a view that users can query by using the Search operation\cr
\link[paws.application.integration:resourceexplorer_delete_index]{delete_index} \tab Deletes the specified index and turns off Amazon Web Services Resource Explorer in the specified Amazon Web Services Region\cr
\link[paws.application.integration:resourceexplorer_delete_resource_explorer_setup]{delete_resource_explorer_setup} \tab Deletes a Resource Explorer setup configuration\cr
\link[paws.application.integration:resourceexplorer_delete_view]{delete_view} \tab Deletes the specified view\cr
\link[paws.application.integration:resourceexplorer_disassociate_default_view]{disassociate_default_view} \tab After you call this operation, the affected Amazon Web Services Region no longer has a default view\cr
\link[paws.application.integration:resourceexplorer_get_account_level_service_configuration]{get_account_level_service_configuration} \tab Retrieves the status of your account's Amazon Web Services service access, and validates the service linked role required to access the multi-account search feature\cr
\link[paws.application.integration:resourceexplorer_get_default_view]{get_default_view} \tab Retrieves the Amazon Resource Name (ARN) of the view that is the default for the Amazon Web Services Region in which you call this operation\cr
\link[paws.application.integration:resourceexplorer_get_index]{get_index} \tab Retrieves details about the Amazon Web Services Resource Explorer index in the Amazon Web Services Region in which you invoked the operation\cr
\link[paws.application.integration:resourceexplorer_get_managed_view]{get_managed_view} \tab Retrieves details of the specified Amazon Web Services-managed view\cr
\link[paws.application.integration:resourceexplorer_get_resource_explorer_setup]{get_resource_explorer_setup} \tab Retrieves the status and details of a Resource Explorer setup operation\cr
\link[paws.application.integration:resourceexplorer_get_service_index]{get_service_index} \tab Retrieves information about the Resource Explorer index in the current Amazon Web Services Region\cr
\link[paws.application.integration:resourceexplorer_get_service_view]{get_service_view} \tab Retrieves details about a specific Resource Explorer service view\cr
\link[paws.application.integration:resourceexplorer_get_view]{get_view} \tab Retrieves details of the specified view\cr
\link[paws.application.integration:resourceexplorer_list_indexes]{list_indexes} \tab Retrieves a list of all of the indexes in Amazon Web Services Regions that are currently collecting resource information for Amazon Web Services Resource Explorer\cr
\link[paws.application.integration:resourceexplorer_list_indexes_for_members]{list_indexes_for_members} \tab Retrieves a list of a member's indexes in all Amazon Web Services Regions that are currently collecting resource information for Amazon Web Services Resource Explorer\cr
\link[paws.application.integration:resourceexplorer_list_managed_views]{list_managed_views} \tab Lists the Amazon resource names (ARNs) of the Amazon Web Services-managed views available in the Amazon Web Services Region in which you call this operation\cr
\link[paws.application.integration:resourceexplorer_list_resources]{list_resources} \tab Returns a list of resources and their details that match the specified criteria\cr
\link[paws.application.integration:resourceexplorer_list_service_indexes]{list_service_indexes} \tab Lists all Resource Explorer indexes across the specified Amazon Web Services Regions\cr
\link[paws.application.integration:resourceexplorer_list_service_views]{list_service_views} \tab Lists all Resource Explorer service views available in the current Amazon Web Services account\cr
\link[paws.application.integration:resourceexplorer_list_streaming_access_for_services]{list_streaming_access_for_services} \tab Returns a list of Amazon Web Services services that have been granted streaming access to your Resource Explorer data\cr
\link[paws.application.integration:resourceexplorer_list_supported_resource_types]{list_supported_resource_types} \tab Retrieves a list of all resource types currently supported by Amazon Web Services Resource Explorer\cr
\link[paws.application.integration:resourceexplorer_list_tags_for_resource]{list_tags_for_resource} \tab Lists the tags that are attached to the specified resource\cr
\link[paws.application.integration:resourceexplorer_list_views]{list_views} \tab Lists the Amazon resource names (ARNs) of the views available in the Amazon Web Services Region in which you call this operation\cr
\link[paws.application.integration:resourceexplorer_search]{search} \tab Searches for resources and displays details about all resources that match the specified criteria\cr
\link[paws.application.integration:resourceexplorer_tag_resource]{tag_resource} \tab Adds one or more tag key and value pairs to an Amazon Web Services Resource Explorer view or index\cr
\link[paws.application.integration:resourceexplorer_untag_resource]{untag_resource} \tab Removes one or more tag key and value pairs from an Amazon Web Services Resource Explorer view or index\cr
\link[paws.application.integration:resourceexplorer_update_index_type]{update_index_type} \tab Changes the type of the index from one of the following types to the other\cr
\link[paws.application.integration:resourceexplorer_update_view]{update_view} \tab Modifies some of the details of a view
}
}

\examples{
\dontrun{
svc <- resourceexplorer()
svc$associate_default_view(
  Foo = 123
)
}

}
