% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_tag}
\alias{resourcegroups_tag}
\title{Adds specified tags to a resource with the specified ARN}
\usage{
resourcegroups_tag(Arn, Tags)
}
\arguments{
\item{Arn}{[required] The ARN of the resource to which to add tags.}

\item{Tags}{[required] The tags to add to the specified resource. A tag is a string-to-string
map of key-value pairs. Tag keys can have a maximum character length of
128 characters, and tag values can have a maximum length of 256
characters.}
}
\description{
Adds specified tags to a resource with the specified ARN. Existing tags
on a resource are not changed if they are not specified in the request
parameters.
}
\section{Request syntax}{
\preformatted{resourcegroups$tag(
  Arn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
