% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_create_robot_application}
\alias{robomaker_create_robot_application}
\title{Creates a robot application}
\usage{
robomaker_create_robot_application(name, sources, robotSoftwareSuite,
  tags)
}
\arguments{
\item{name}{[required] The name of the robot application.}

\item{sources}{[required] The sources of the robot application.}

\item{robotSoftwareSuite}{[required] The robot software suite used by the robot application.}

\item{tags}{A map that contains tag keys and tag values that are attached to the
robot application.}
}
\description{
Creates a robot application.
}
\section{Request syntax}{
\preformatted{svc$create_robot_application(
  name = "string",
  sources = list(
    list(
      s3Bucket = "string",
      s3Key = "string",
      architecture = "X86_64"|"ARM64"|"ARMHF"
    )
  ),
  robotSoftwareSuite = list(
    name = "ROS",
    version = "Kinetic"
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
