% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_deployment_job}
\alias{robomaker_describe_deployment_job}
\title{Describes a deployment job}
\usage{
robomaker_describe_deployment_job(job)
}
\arguments{
\item{job}{[required] The Amazon Resource Name (ARN) of the deployment job.}
}
\description{
Describes a deployment job.
}
\section{Request syntax}{
\preformatted{svc$describe_deployment_job(
  job = "string"
)
}
}

\keyword{internal}
