% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_fleet}
\alias{robomaker_describe_fleet}
\title{Describes a fleet}
\usage{
robomaker_describe_fleet(fleet)
}
\arguments{
\item{fleet}{[required] The Amazon Resource Name (ARN) of the fleet.}
}
\description{
Describes a fleet.
}
\section{Request syntax}{
\preformatted{robomaker$describe_fleet(
  fleet = "string"
)
}
}

\keyword{internal}
