% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_robot}
\alias{robomaker_describe_robot}
\title{Describes a robot}
\usage{
robomaker_describe_robot(robot)
}
\arguments{
\item{robot}{[required] The Amazon Resource Name (ARN) of the robot to be described.}
}
\description{
Describes a robot.
}
\section{Request syntax}{
\preformatted{robomaker$describe_robot(
  robot = "string"
)
}
}

\keyword{internal}
