% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_robot_application}
\alias{robomaker_describe_robot_application}
\title{Describes a robot application}
\usage{
robomaker_describe_robot_application(application, applicationVersion)
}
\arguments{
\item{application}{[required] The Amazon Resource Name (ARN) of the robot application.}

\item{applicationVersion}{The version of the robot application to describe.}
}
\description{
Describes a robot application.
}
\section{Request syntax}{
\preformatted{svc$describe_robot_application(
  application = "string",
  applicationVersion = "string"
)
}
}

\keyword{internal}
