% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_simulation_job}
\alias{robomaker_describe_simulation_job}
\title{Describes a simulation job}
\usage{
robomaker_describe_simulation_job(job)
}
\arguments{
\item{job}{[required] The Amazon Resource Name (ARN) of the simulation job to be described.}
}
\description{
Describes a simulation job.
}
\section{Request syntax}{
\preformatted{robomaker$describe_simulation_job(
  job = "string"
)
}
}

\keyword{internal}
