% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_world}
\alias{robomaker_describe_world}
\title{Describes a world}
\usage{
robomaker_describe_world(world)
}
\arguments{
\item{world}{[required] The Amazon Resource Name (arn) of the world you want to describe.}
}
\description{
Describes a world.
}
\section{Request syntax}{
\preformatted{svc$describe_world(
  world = "string"
)
}
}

\keyword{internal}
