% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_world_template}
\alias{robomaker_describe_world_template}
\title{Describes a world template}
\usage{
robomaker_describe_world_template(template)
}
\arguments{
\item{template}{[required] The Amazon Resource Name (arn) of the world template you want to
describe.}
}
\description{
Describes a world template.
}
\section{Request syntax}{
\preformatted{svc$describe_world_template(
  template = "string"
)
}
}

\keyword{internal}
