% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_deployment_jobs}
\alias{robomaker_list_deployment_jobs}
\title{Returns a list of deployment jobs for a fleet}
\usage{
robomaker_list_deployment_jobs(filters, nextToken, maxResults)
}
\arguments{
\item{filters}{Optional filters to limit results.

The filter names \code{status} and \code{fleetName} are supported. When filtering,
you must use the complete value of the filtered item. You can use up to
three filters, but they must be for the same named item. For example, if
you are looking for items with the status \code{InProgress} or the status
\code{Pending}.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{ListDeploymentJobs} request where \code{maxResults} was used and the results
exceeded the value of that parameter. Pagination continues from the end
of the previous results that returned the \code{nextToken} value.

This token should be treated as an opaque identifier that is only used
to retrieve the next items in a list and not for other programmatic
purposes.}

\item{maxResults}{The maximum number of deployment job results returned by
\code{ListDeploymentJobs} in paginated output. When this parameter is used,
\code{ListDeploymentJobs} only returns \code{maxResults} results in a single page
along with a \code{nextToken} response element. The remaining results of the
initial request can be seen by sending another \code{ListDeploymentJobs}
request with the returned \code{nextToken} value. This value can be between 1
and 100. If this parameter is not used, then \code{ListDeploymentJobs}
returns up to 100 results and a \code{nextToken} value if applicable.}
}
\description{
Returns a list of deployment jobs for a fleet. You can optionally
provide filters to retrieve specific deployment jobs.
}
\section{Request syntax}{
\preformatted{svc$list_deployment_jobs(
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
