% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_robot_applications}
\alias{robomaker_list_robot_applications}
\title{Returns a list of robot application}
\usage{
robomaker_list_robot_applications(versionQualifier, nextToken,
  maxResults, filters)
}
\arguments{
\item{versionQualifier}{The version qualifier of the robot application.}

\item{nextToken}{If the previous paginated request did not return all of the remaining
results, the response object's \code{nextToken} parameter value is set to a
token. To retrieve the next set of results, call
\code{\link[=robomaker_list_robot_applications]{list_robot_applications}} again and
assign that token to the request object's \code{nextToken} parameter. If
there are no remaining results, the previous response object's NextToken
parameter is set to null.}

\item{maxResults}{When this parameter is used,
\code{\link[=robomaker_list_robot_applications]{list_robot_applications}} only
returns \code{maxResults} results in a single page along with a \code{nextToken}
response element. The remaining results of the initial request can be
seen by sending another
\code{\link[=robomaker_list_robot_applications]{list_robot_applications}} request
with the returned \code{nextToken} value. This value can be between 1 and
100. If this parameter is not used, then
\code{\link[=robomaker_list_robot_applications]{list_robot_applications}} returns
up to 100 results and a \code{nextToken} value if applicable.}

\item{filters}{Optional filters to limit results.

The filter name \code{name} is supported. When filtering, you must use the
complete value of the filtered item. You can use up to three filters.}
}
\description{
Returns a list of robot application. You can optionally provide filters
to retrieve specific robot applications.
}
\section{Request syntax}{
\preformatted{svc$list_robot_applications(
  versionQualifier = "string",
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
