% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_simulation_job_batches}
\alias{robomaker_list_simulation_job_batches}
\title{Returns a list simulation job batches}
\usage{
robomaker_list_simulation_job_batches(nextToken, maxResults, filters)
}
\arguments{
\item{nextToken}{If the previous paginated request did not return all of the remaining
results, the response object's \code{nextToken} parameter value is set to a
token. To retrieve the next set of results, call
\code{ListSimulationJobBatches} again and assign that token to the request
object's \code{nextToken} parameter. If there are no remaining results, the
previous response object's NextToken parameter is set to null.}

\item{maxResults}{When this parameter is used, \code{ListSimulationJobBatches} only returns
\code{maxResults} results in a single page along with a \code{nextToken} response
element. The remaining results of the initial request can be seen by
sending another \code{ListSimulationJobBatches} request with the returned
\code{nextToken} value.}

\item{filters}{Optional filters to limit results.}
}
\description{
Returns a list simulation job batches. You can optionally provide
filters to retrieve specific simulation batch jobs.
}
\section{Request syntax}{
\preformatted{svc$list_simulation_job_batches(
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
