% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_register_robot}
\alias{robomaker_register_robot}
\title{Registers a robot with a fleet}
\usage{
robomaker_register_robot(fleet, robot)
}
\arguments{
\item{fleet}{[required] The Amazon Resource Name (ARN) of the fleet.}

\item{robot}{[required] The Amazon Resource Name (ARN) of the robot.}
}
\description{
Registers a robot with a fleet.
}
\section{Request syntax}{
\preformatted{robomaker$register_robot(
  fleet = "string",
  robot = "string"
)
}
}

\keyword{internal}
