% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_start_simulation_job_batch}
\alias{robomaker_start_simulation_job_batch}
\title{Starts a new simulation job batch}
\usage{
robomaker_start_simulation_job_batch(clientRequestToken, batchPolicy,
  createSimulationJobRequests, tags)
}
\arguments{
\item{clientRequestToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{batchPolicy}{The batch policy.}

\item{createSimulationJobRequests}{[required] A list of simulation job requests to create in the batch.}

\item{tags}{A map that contains tag keys and tag values that are attached to the
deployment job batch.}
}
\description{
Starts a new simulation job batch. The batch is defined using one or
more \code{SimulationJobRequest} objects.
}
\section{Request syntax}{
\preformatted{svc$start_simulation_job_batch(
  clientRequestToken = "string",
  batchPolicy = list(
    timeoutInSeconds = 123,
    maxConcurrency = 123
  ),
  createSimulationJobRequests = list(
    list(
      outputLocation = list(
        s3Bucket = "string",
        s3Prefix = "string"
      ),
      loggingConfig = list(
        recordAllRosTopics = TRUE|FALSE
      ),
      maxJobDurationInSeconds = 123,
      iamRole = "string",
      failureBehavior = "Fail"|"Continue",
      useDefaultApplications = TRUE|FALSE,
      robotApplications = list(
        list(
          application = "string",
          applicationVersion = "string",
          launchConfig = list(
            packageName = "string",
            launchFile = "string",
            environmentVariables = list(
              "string"
            ),
            portForwardingConfig = list(
              portMappings = list(
                list(
                  jobPort = 123,
                  applicationPort = 123,
                  enableOnPublicIp = TRUE|FALSE
                )
              )
            ),
            streamUI = TRUE|FALSE
          )
        )
      ),
      simulationApplications = list(
        list(
          application = "string",
          applicationVersion = "string",
          launchConfig = list(
            packageName = "string",
            launchFile = "string",
            environmentVariables = list(
              "string"
            ),
            portForwardingConfig = list(
              portMappings = list(
                list(
                  jobPort = 123,
                  applicationPort = 123,
                  enableOnPublicIp = TRUE|FALSE
                )
              )
            ),
            streamUI = TRUE|FALSE
          )
        )
      ),
      dataSources = list(
        list(
          name = "string",
          s3Bucket = "string",
          s3Keys = list(
            "string"
          )
        )
      ),
      vpcConfig = list(
        subnets = list(
          "string"
        ),
        securityGroups = list(
          "string"
        ),
        assignPublicIp = TRUE|FALSE
      ),
      compute = list(
        simulationUnitLimit = 123
      ),
      tags = list(
        "string"
      )
    )
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
