% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_delete_reusable_delegation_set}
\alias{route53_delete_reusable_delegation_set}
\title{Deletes a reusable delegation set}
\usage{
route53_delete_reusable_delegation_set(Id)
}
\arguments{
\item{Id}{[required] The ID of the reusable delegation set that you want to delete.}
}
\description{
Deletes a reusable delegation set.
}
\details{
You can delete a reusable delegation set only if it isn't associated
with any hosted zones.

To verify that the reusable delegation set is not associated with any
hosted zones, submit a
\href{https://docs.aws.amazon.com/Route53/latest/APIReference/API_GetReusableDelegationSet.html}{GetReusableDelegationSet}
request and specify the ID of the reusable delegation set that you want
to delete.
}
\section{Request syntax}{
\preformatted{svc$delete_reusable_delegation_set(
  Id = "string"
)
}
}

\keyword{internal}
