% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_dnssec}
\alias{route53_get_dnssec}
\title{Returns information about DNSSEC for a specific hosted zone, including
the key signing keys (KSKs) and zone signing keys (ZSKs) in the hosted
zone}
\usage{
route53_get_dnssec(HostedZoneId)
}
\arguments{
\item{HostedZoneId}{[required] A unique string used to identify a hosted zone.}
}
\description{
Returns information about DNSSEC for a specific hosted zone, including
the key signing keys (KSKs) and zone signing keys (ZSKs) in the hosted
zone.
}
\section{Request syntax}{
\preformatted{svc$get_dnssec(
  HostedZoneId = "string"
)
}
}

\keyword{internal}
