% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_hosted_zone}
\alias{route53_get_hosted_zone}
\title{Gets information about a specified hosted zone including the four name
servers assigned to the hosted zone}
\usage{
route53_get_hosted_zone(Id)
}
\arguments{
\item{Id}{[required] The ID of the hosted zone that you want to get information about.}
}
\description{
Gets information about a specified hosted zone including the four name
servers assigned to the hosted zone.
}
\section{Request syntax}{
\preformatted{svc$get_hosted_zone(
  Id = "string"
)
}
}

\examples{
# The following example gets information about the Z3M3LMPEXAMPLE hosted
# zone.
\donttest{svc$get_hosted_zone(
  Id = "Z3M3LMPEXAMPLE"
)}

}
\keyword{internal}
