% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_disable_domain_auto_renew}
\alias{route53domains_disable_domain_auto_renew}
\title{This operation disables automatic renewal of domain registration for the
specified domain}
\usage{
route53domains_disable_domain_auto_renew(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to disable automatic renewal for.}
}
\description{
This operation disables automatic renewal of domain registration for the
specified domain.
}
\section{Request syntax}{
\preformatted{route53domains$disable_domain_auto_renew(
  DomainName = "string"
)
}
}

\keyword{internal}
