% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_resend_contact_reachability_email}
\alias{route53domains_resend_contact_reachability_email}
\title{For operations that require confirmation that the email address for the
registrant contact is valid, such as registering a new domain, this
operation resends the confirmation email to the current email address
for the registrant contact}
\usage{
route53domains_resend_contact_reachability_email(domainName)
}
\arguments{
\item{domainName}{The name of the domain for which you want Amazon Route 53 to resend a
confirmation email to the registrant contact.}
}
\description{
For operations that require confirmation that the email address for the
registrant contact is valid, such as registering a new domain, this
operation resends the confirmation email to the current email address
for the registrant contact.
}
\section{Request syntax}{
\preformatted{route53domains$resend_contact_reachability_email(
  domainName = "string"
)
}
}

\keyword{internal}
