% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_retrieve_domain_auth_code}
\alias{route53domains_retrieve_domain_auth_code}
\title{This operation returns the AuthCode for the domain}
\usage{
route53domains_retrieve_domain_auth_code(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to get an authorization code for.}
}
\description{
This operation returns the AuthCode for the domain. To transfer a domain
to another registrar, you provide this value to the new registrar.
}
\section{Request syntax}{
\preformatted{route53domains$retrieve_domain_auth_code(
  DomainName = "string"
)
}
}

\keyword{internal}
