% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_delete_resolver_rule}
\alias{route53resolver_delete_resolver_rule}
\title{Deletes a resolver rule}
\usage{
route53resolver_delete_resolver_rule(ResolverRuleId)
}
\arguments{
\item{ResolverRuleId}{[required] The ID of the resolver rule that you want to delete.}
}
\description{
Deletes a resolver rule. Before you can delete a resolver rule, you must
disassociate it from all the VPCs that you associated the resolver rule
with. For more infomation, see DisassociateResolverRule.
}
\section{Request syntax}{
\preformatted{route53resolver$delete_resolver_rule(
  ResolverRuleId = "string"
)
}
}

\keyword{internal}
