% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_resolver_rule_association}
\alias{route53resolver_get_resolver_rule_association}
\title{Gets information about an association between a specified resolver rule
and a VPC}
\usage{
route53resolver_get_resolver_rule_association(ResolverRuleAssociationId)
}
\arguments{
\item{ResolverRuleAssociationId}{[required] The ID of the resolver rule association that you want to get information
about.}
}
\description{
Gets information about an association between a specified resolver rule
and a VPC. You associate a resolver rule and a VPC using
AssociateResolverRule.
}
\section{Request syntax}{
\preformatted{route53resolver$get_resolver_rule_association(
  ResolverRuleAssociationId = "string"
)
}
}

\keyword{internal}
