% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_encryption}
\alias{s3_delete_bucket_encryption}
\title{This implementation of the DELETE operation removes default encryption
from the bucket}
\usage{
s3_delete_bucket_encryption(Bucket, ExpectedBucketOwner)
}
\arguments{
\item{Bucket}{[required] The name of the bucket containing the server-side encryption
configuration to delete.}

\item{ExpectedBucketOwner}{The account id of the expected bucket owner. If the bucket is owned by a
different account, the request will fail with an HTTP
\verb{403 (Access Denied)} error.}
}
\description{
This implementation of the DELETE operation removes default encryption
from the bucket. For information about the Amazon S3 default encryption
feature, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/bucket-encryption.html}{Amazon S3 Default Bucket Encryption}
in the \emph{Amazon Simple Storage Service Developer Guide}.

To use this operation, you must have permissions to perform the
\code{s3:PutEncryptionConfiguration} action. The bucket owner has this
permission by default. The bucket owner can grant this permission to
others. For more information about permissions, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/using-with-s3-actions.html#using-with-s3-actions-related-to-bucket-subresources}{Permissions Related to Bucket Subresource Operations}
and \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-access-control.html}{Managing Access Permissions to your Amazon S3 Resources}
in the \emph{Amazon Simple Storage Service Developer Guide}.

\strong{Related Resources}
\itemize{
\item \code{\link[=s3_put_bucket_encryption]{put_bucket_encryption}}
\item \code{\link[=s3_get_bucket_encryption]{get_bucket_encryption}}
}
}
\section{Request syntax}{
\preformatted{svc$delete_bucket_encryption(
  Bucket = "string",
  ExpectedBucketOwner = "string"
)
}
}

\keyword{internal}
