% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_metrics_configuration}
\alias{s3_delete_bucket_metrics_configuration}
\title{Deletes a metrics configuration (specified by the metrics configuration
ID) from the bucket}
\usage{
s3_delete_bucket_metrics_configuration(Bucket, Id)
}
\arguments{
\item{Bucket}{[required] The name of the bucket containing the metrics configuration to delete.}

\item{Id}{[required] The ID used to identify the metrics configuration.}
}
\description{
Deletes a metrics configuration (specified by the metrics configuration
ID) from the bucket.
}
\section{Request syntax}{
\preformatted{s3$delete_bucket_metrics_configuration(
  Bucket = "string",
  Id = "string"
)
}
}

\keyword{internal}
