% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_accelerate_configuration}
\alias{s3_get_bucket_accelerate_configuration}
\title{Returns the accelerate configuration of a bucket}
\usage{
s3_get_bucket_accelerate_configuration(Bucket)
}
\arguments{
\item{Bucket}{[required] Name of the bucket for which the accelerate configuration is retrieved.}
}
\description{
Returns the accelerate configuration of a bucket.
}
\section{Request syntax}{
\preformatted{svc$get_bucket_accelerate_configuration(
  Bucket = "string"
)
}
}

\keyword{internal}
