% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_notification_configuration}
\alias{s3_get_bucket_notification_configuration}
\title{Returns the notification configuration of a bucket}
\usage{
s3_get_bucket_notification_configuration(Bucket)
}
\arguments{
\item{Bucket}{[required] Name of the bucket to get the notification configuration for.}
}
\description{
Returns the notification configuration of a bucket.
}
\section{Request syntax}{
\preformatted{svc$get_bucket_notification_configuration(
  Bucket = "string"
)
}
}

\keyword{internal}
