% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_website}
\alias{s3_get_bucket_website}
\title{Returns the website configuration for a bucket}
\usage{
s3_get_bucket_website(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the website configuration for a bucket.
}
\section{Request syntax}{
\preformatted{s3$get_bucket_website(
  Bucket = "string"
)
}
}

\examples{
# The following example retrieves website configuration of a bucket.
\donttest{s3$get_bucket_website(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
