% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_object_legal_hold}
\alias{s3_get_object_legal_hold}
\title{Gets an object's current Legal Hold status}
\usage{
s3_get_object_legal_hold(Bucket, Key, VersionId, RequestPayer)
}
\arguments{
\item{Bucket}{[required] The bucket containing the object whose Legal Hold status you want to
retrieve.}

\item{Key}{[required] The key name for the object whose Legal Hold status you want to
retrieve.}

\item{VersionId}{The version ID of the object whose Legal Hold status you want to
retrieve.}

\item{RequestPayer}{}
}
\description{
Gets an object's current Legal Hold status.
}
\section{Request syntax}{
\preformatted{s3$get_object_legal_hold(
  Bucket = "string",
  Key = "string",
  VersionId = "string",
  RequestPayer = "requester"
)
}
}

\keyword{internal}
