% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_bucket_accelerate_configuration}
\alias{s3_put_bucket_accelerate_configuration}
\title{Sets the accelerate configuration of an existing bucket}
\usage{
s3_put_bucket_accelerate_configuration(Bucket, AccelerateConfiguration)
}
\arguments{
\item{Bucket}{[required] Name of the bucket for which the accelerate configuration is set.}

\item{AccelerateConfiguration}{[required] Specifies the Accelerate Configuration you want to set for the bucket.}
}
\description{
Sets the accelerate configuration of an existing bucket.
}
\section{Request syntax}{
\preformatted{s3$put_bucket_accelerate_configuration(
  Bucket = "string",
  AccelerateConfiguration = list(
    Status = "Enabled"|"Suspended"
  )
)
}
}

\keyword{internal}
