% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_describe_job}
\alias{s3control_describe_job}
\title{Retrieves the configuration parameters and status for a batch operations
job}
\usage{
s3control_describe_job(AccountId, JobId)
}
\arguments{
\item{AccountId}{[required]}

\item{JobId}{[required] The ID for the job whose information you want to retrieve.}
}
\description{
Retrieves the configuration parameters and status for a batch operations
job.
}
\section{Request syntax}{
\preformatted{svc$describe_job(
  AccountId = "string",
  JobId = "string"
)
}
}

\keyword{internal}
