% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_public_access_block}
\alias{s3control_get_public_access_block}
\title{Get public access block}
\usage{
s3control_get_public_access_block(AccountId)
}
\arguments{
\item{AccountId}{[required]}
}
\description{
Get public access block
}
\section{Request syntax}{
\preformatted{svc$get_public_access_block(
  AccountId = "string"
)
}
}

\keyword{internal}
