% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3outposts_operations.R
\name{s3outposts_list_endpoints}
\alias{s3outposts_list_endpoints}
\title{S3 on Outposts access points simplify managing data access at scale for
shared datasets in Amazon S3 on Outposts}
\usage{
s3outposts_list_endpoints(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The next endpoint requested in the list.}

\item{MaxResults}{The max number of endpoints that can be returned on the request.}
}
\description{
S3 on Outposts access points simplify managing data access at scale for
shared datasets in Amazon S3 on Outposts. S3 on Outposts uses endpoints
to connect to Outposts buckets so that you can perform actions within
your virtual private cloud (VPC).

This action lists endpoints associated with the Outpost.

Related actions include:
\itemize{
\item \href{https://docs.aws.amazon.com/AmazonS3/latest/API/API_s3outposts_CreateEndpoint.html}{CreateEndpoint}
\item \href{https://docs.aws.amazon.com/AmazonS3/latest/API/API_s3outposts_DeleteEndpoint.html}{DeleteEndpoint}
}
}
\section{Request syntax}{
\preformatted{svc$list_endpoints(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
