% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_associate_trial_component}
\alias{sagemaker_associate_trial_component}
\title{Associates a trial component with a trial}
\usage{
sagemaker_associate_trial_component(TrialComponentName, TrialName)
}
\arguments{
\item{TrialComponentName}{[required] The name of the component to associated with the trial.}

\item{TrialName}{[required] The name of the trial to associate with.}
}
\description{
Associates a trial component with a trial. A trial component can be
associated with multiple trials. To disassociate a trial component from
a trial, call the DisassociateTrialComponent API.
}
\section{Request syntax}{
\preformatted{svc$associate_trial_component(
  TrialComponentName = "string",
  TrialName = "string"
)
}
}

\keyword{internal}
