% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_monitoring_schedule}
\alias{sagemaker_delete_monitoring_schedule}
\title{Deletes a monitoring schedule}
\usage{
sagemaker_delete_monitoring_schedule(MonitoringScheduleName)
}
\arguments{
\item{MonitoringScheduleName}{[required] The name of the monitoring schedule to delete.}
}
\description{
Deletes a monitoring schedule. Also stops the schedule had not already
been stopped. This does not delete the job execution history of the
monitoring schedule.
}
\section{Request syntax}{
\preformatted{svc$delete_monitoring_schedule(
  MonitoringScheduleName = "string"
)
}
}

\keyword{internal}
