% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_notebook_instance_lifecycle_config}
\alias{sagemaker_delete_notebook_instance_lifecycle_config}
\title{Deletes a notebook instance lifecycle configuration}
\usage{
sagemaker_delete_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName)
}
\arguments{
\item{NotebookInstanceLifecycleConfigName}{[required] The name of the lifecycle configuration to delete.}
}
\description{
Deletes a notebook instance lifecycle configuration.
}
\section{Request syntax}{
\preformatted{sagemaker$delete_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName = "string"
)
}
}

\keyword{internal}
