% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_action}
\alias{sagemaker_describe_action}
\title{Describes an action}
\usage{
sagemaker_describe_action(ActionName)
}
\arguments{
\item{ActionName}{[required] The name of the action to describe.}
}
\description{
Describes an action.
}
\section{Request syntax}{
\preformatted{svc$describe_action(
  ActionName = "string"
)
}
}

\keyword{internal}
