% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_algorithm}
\alias{sagemaker_describe_algorithm}
\title{Returns a description of the specified algorithm that is in your account}
\usage{
sagemaker_describe_algorithm(AlgorithmName)
}
\arguments{
\item{AlgorithmName}{[required] The name of the algorithm to describe.}
}
\description{
Returns a description of the specified algorithm that is in your
account.
}
\section{Request syntax}{
\preformatted{svc$describe_algorithm(
  AlgorithmName = "string"
)
}
}

\keyword{internal}
