% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_auto_ml_job}
\alias{sagemaker_describe_auto_ml_job}
\title{Returns information about an Amazon SageMaker job}
\usage{
sagemaker_describe_auto_ml_job(AutoMLJobName)
}
\arguments{
\item{AutoMLJobName}{[required] Request information about a job using that job\'s unique name.}
}
\description{
Returns information about an Amazon SageMaker job.
}
\section{Request syntax}{
\preformatted{svc$describe_auto_ml_job(
  AutoMLJobName = "string"
)
}
}

\keyword{internal}
