% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_context}
\alias{sagemaker_describe_context}
\title{Describes a context}
\usage{
sagemaker_describe_context(ContextName)
}
\arguments{
\item{ContextName}{[required] The name of the context to describe.}
}
\description{
Describes a context.
}
\section{Request syntax}{
\preformatted{svc$describe_context(
  ContextName = "string"
)
}
}

\keyword{internal}
