% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_domain}
\alias{sagemaker_describe_domain}
\title{The desciption of the domain}
\usage{
sagemaker_describe_domain(DomainId)
}
\arguments{
\item{DomainId}{[required] The domain ID.}
}
\description{
The desciption of the domain.
}
\section{Request syntax}{
\preformatted{svc$describe_domain(
  DomainId = "string"
)
}
}

\keyword{internal}
