% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_flow_definition}
\alias{sagemaker_describe_flow_definition}
\title{Returns information about the specified flow definition}
\usage{
sagemaker_describe_flow_definition(FlowDefinitionName)
}
\arguments{
\item{FlowDefinitionName}{[required] The name of the flow definition.}
}
\description{
Returns information about the specified flow definition.
}
\section{Request syntax}{
\preformatted{svc$describe_flow_definition(
  FlowDefinitionName = "string"
)
}
}

\keyword{internal}
