% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_model_bias_job_definition}
\alias{sagemaker_describe_model_bias_job_definition}
\title{Returns a description of a model bias job definition}
\usage{
sagemaker_describe_model_bias_job_definition(JobDefinitionName)
}
\arguments{
\item{JobDefinitionName}{[required] The name of the model bias job definition. The name must be unique
within an AWS Region in the AWS account.}
}
\description{
Returns a description of a model bias job definition.
}
\section{Request syntax}{
\preformatted{svc$describe_model_bias_job_definition(
  JobDefinitionName = "string"
)
}
}

\keyword{internal}
