% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_notebook_instance_lifecycle_config}
\alias{sagemaker_describe_notebook_instance_lifecycle_config}
\title{Returns a description of a notebook instance lifecycle configuration}
\usage{
sagemaker_describe_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName)
}
\arguments{
\item{NotebookInstanceLifecycleConfigName}{[required] The name of the lifecycle configuration to describe.}
}
\description{
Returns a description of a notebook instance lifecycle configuration.
}
\details{
For information about notebook instance lifestyle configurations, see
\href{https://docs.aws.amazon.com/sagemaker/latest/dg/notebook-lifecycle-config.html}{Step 2.1: (Optional) Customize a Notebook Instance}.
}
\section{Request syntax}{
\preformatted{sagemaker$describe_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName = "string"
)
}
}

\keyword{internal}
