% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_processing_job}
\alias{sagemaker_describe_processing_job}
\title{Returns a description of a processing job}
\usage{
sagemaker_describe_processing_job(ProcessingJobName)
}
\arguments{
\item{ProcessingJobName}{[required] The name of the processing job. The name must be unique within an AWS
Region in the AWS account.}
}
\description{
Returns a description of a processing job.
}
\section{Request syntax}{
\preformatted{svc$describe_processing_job(
  ProcessingJobName = "string"
)
}
}

\keyword{internal}
