% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_project}
\alias{sagemaker_describe_project}
\title{Describes the details of a project}
\usage{
sagemaker_describe_project(ProjectName)
}
\arguments{
\item{ProjectName}{[required] The name of the project to describe.}
}
\description{
Describes the details of a project.
}
\section{Request syntax}{
\preformatted{svc$describe_project(
  ProjectName = "string"
)
}
}

\keyword{internal}
