% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_trial}
\alias{sagemaker_describe_trial}
\title{Provides a list of a trial's properties}
\usage{
sagemaker_describe_trial(TrialName)
}
\arguments{
\item{TrialName}{[required] The name of the trial to describe.}
}
\description{
Provides a list of a trial's properties.
}
\section{Request syntax}{
\preformatted{svc$describe_trial(
  TrialName = "string"
)
}
}

\keyword{internal}
