% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_get_device_fleet_report}
\alias{sagemaker_get_device_fleet_report}
\title{Describes a fleet}
\usage{
sagemaker_get_device_fleet_report(DeviceFleetName)
}
\arguments{
\item{DeviceFleetName}{[required] The name of the fleet.}
}
\description{
Describes a fleet.
}
\section{Request syntax}{
\preformatted{svc$get_device_fleet_report(
  DeviceFleetName = "string"
)
}
}

\keyword{internal}
