% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_domains}
\alias{sagemaker_list_domains}
\title{Lists the domains}
\usage{
sagemaker_list_domains(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{If the previous response was truncated, you will receive this token. Use
it in your next request to receive the next set of results.}

\item{MaxResults}{Returns a list up to a specified limit.}
}
\description{
Lists the domains.
}
\section{Request syntax}{
\preformatted{svc$list_domains(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
