% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_register_devices}
\alias{sagemaker_register_devices}
\title{Register devices}
\usage{
sagemaker_register_devices(DeviceFleetName, Devices, Tags)
}
\arguments{
\item{DeviceFleetName}{[required] The name of the fleet.}

\item{Devices}{[required] A list of devices to register with SageMaker Edge Manager.}

\item{Tags}{The tags associated with devices.}
}
\description{
Register devices.
}
\section{Request syntax}{
\preformatted{svc$register_devices(
  DeviceFleetName = "string",
  Devices = list(
    list(
      DeviceName = "string",
      Description = "string",
      IotThingName = "string"
    )
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
