% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_start_notebook_instance}
\alias{sagemaker_start_notebook_instance}
\title{Launches an ML compute instance with the latest version of the libraries
and attaches your ML storage volume}
\usage{
sagemaker_start_notebook_instance(NotebookInstanceName)
}
\arguments{
\item{NotebookInstanceName}{[required] The name of the notebook instance to start.}
}
\description{
Launches an ML compute instance with the latest version of the libraries
and attaches your ML storage volume. After configuring the notebook
instance, Amazon SageMaker sets the notebook instance status to
\code{InService}. A notebook instance's status must be \code{InService} before
you can connect to your Jupyter notebook.
}
\section{Request syntax}{
\preformatted{svc$start_notebook_instance(
  NotebookInstanceName = "string"
)
}
}

\keyword{internal}
