% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_processing_job}
\alias{sagemaker_stop_processing_job}
\title{Stops a processing job}
\usage{
sagemaker_stop_processing_job(ProcessingJobName)
}
\arguments{
\item{ProcessingJobName}{[required] The name of the processing job to stop.}
}
\description{
Stops a processing job.
}
\section{Request syntax}{
\preformatted{svc$stop_processing_job(
  ProcessingJobName = "string"
)
}
}

\keyword{internal}
