% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_device_fleet}
\alias{sagemaker_update_device_fleet}
\title{Updates a fleet of devices}
\usage{
sagemaker_update_device_fleet(DeviceFleetName, RoleArn, Description,
  OutputConfig)
}
\arguments{
\item{DeviceFleetName}{[required] The name of the fleet.}

\item{RoleArn}{The Amazon Resource Name (ARN) of the device.}

\item{Description}{Description of the fleet.}

\item{OutputConfig}{[required] Output configuration for storing sample data collected by the fleet.}
}
\description{
Updates a fleet of devices.
}
\section{Request syntax}{
\preformatted{svc$update_device_fleet(
  DeviceFleetName = "string",
  RoleArn = "string",
  Description = "string",
  OutputConfig = list(
    S3OutputLocation = "string",
    KmsKeyId = "string"
  )
)
}
}

\keyword{internal}
