% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakeredgemanager_operations.R
\name{sagemakeredgemanager_send_heartbeat}
\alias{sagemakeredgemanager_send_heartbeat}
\title{Use to get the current status of devices registered on SageMaker Edge
Manager}
\usage{
sagemakeredgemanager_send_heartbeat(AgentMetrics, Models, AgentVersion,
  DeviceName, DeviceFleetName)
}
\arguments{
\item{AgentMetrics}{For internal use. Returns a list of SageMaker Edge Manager agent
operating metrics.}

\item{Models}{Returns a list of models deployed on the the device.}

\item{AgentVersion}{[required] Returns the version of the agent.}

\item{DeviceName}{[required] The unique name of the device.}

\item{DeviceFleetName}{[required] The name of the fleet that the device belongs to.}
}
\description{
Use to get the current status of devices registered on SageMaker Edge
Manager.
}
\section{Request syntax}{
\preformatted{svc$send_heartbeat(
  AgentMetrics = list(
    list(
      Dimension = "string",
      MetricName = "string",
      Value = 123.0,
      Timestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  Models = list(
    list(
      ModelName = "string",
      ModelVersion = "string",
      LatestSampleTime = as.POSIXct(
        "2015-01-01"
      ),
      LatestInference = as.POSIXct(
        "2015-01-01"
      ),
      ModelMetrics = list(
        list(
          Dimension = "string",
          MetricName = "string",
          Value = 123.0,
          Timestamp = as.POSIXct(
            "2015-01-01"
          )
        )
      )
    )
  ),
  AgentVersion = "string",
  DeviceName = "string",
  DeviceFleetName = "string"
)
}
}

\keyword{internal}
