% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakerfeaturestoreruntime_operations.R
\name{sagemakerfeaturestoreruntime_get_record}
\alias{sagemakerfeaturestoreruntime_get_record}
\title{Use for OnlineStore serving from a FeatureStore}
\usage{
sagemakerfeaturestoreruntime_get_record(FeatureGroupName,
  RecordIdentifierValueAsString, FeatureNames)
}
\arguments{
\item{FeatureGroupName}{[required] The name of the feature group in which you want to put the records.}

\item{RecordIdentifierValueAsString}{[required] The value that corresponds to \code{RecordIdentifier} type and uniquely
identifies the record in the \code{FeatureGroup}.}

\item{FeatureNames}{List of names of Features to be retrieved. If not specified, the latest
value for all the Features are returned.}
}
\description{
Use for \code{OnlineStore} serving from a \code{FeatureStore}. Only the latest
records stored in the \code{OnlineStore} can be retrieved. If no Record with
\code{RecordIdentifierValue} is found, then an empty result is returned.
}
\section{Request syntax}{
\preformatted{svc$get_record(
  FeatureGroupName = "string",
  RecordIdentifierValueAsString = "string",
  FeatureNames = list(
    "string"
  )
)
}
}

\keyword{internal}
