% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_create_savings_plan}
\alias{savingsplans_create_savings_plan}
\title{Creates a Savings Plan}
\usage{
savingsplans_create_savings_plan(savingsPlanOfferingId, commitment,
  upfrontPaymentAmount, clientToken, tags)
}
\arguments{
\item{savingsPlanOfferingId}{[required] The ID of the offering.}

\item{commitment}{[required] The hourly commitment, in USD. This is a value between 0.001 and 1
million. You cannot specify more than three digits after the decimal
point.}

\item{upfrontPaymentAmount}{The up-front payment amount. This is a whole number between 50 and 99
percent of the total value of the Savings Plan. This parameter is
supported only if the payment option is \verb{Partial Upfront}.}

\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{tags}{One or more tags.}
}
\description{
Creates a Savings Plan.
}
\section{Request syntax}{
\preformatted{svc$create_savings_plan(
  savingsPlanOfferingId = "string",
  commitment = "string",
  upfrontPaymentAmount = "string",
  clientToken = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
