% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_delete_queued_savings_plan}
\alias{savingsplans_delete_queued_savings_plan}
\title{Deletes the queued purchase for the specified Savings Plan}
\usage{
savingsplans_delete_queued_savings_plan(savingsPlanId)
}
\arguments{
\item{savingsPlanId}{[required] The ID of the Savings Plan.}
}
\description{
Deletes the queued purchase for the specified Savings Plan.
}
\section{Request syntax}{
\preformatted{svc$delete_queued_savings_plan(
  savingsPlanId = "string"
)
}
}

\keyword{internal}
