% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_describe_savings_plan_rates}
\alias{savingsplans_describe_savings_plan_rates}
\title{Describes the specified Savings Plans rates}
\usage{
savingsplans_describe_savings_plan_rates(savingsPlanId, filters,
  nextToken, maxResults)
}
\arguments{
\item{savingsPlanId}{[required] The ID of the Savings Plan.}

\item{filters}{The filters.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
additional results, make another call with the returned token value.}
}
\description{
Describes the specified Savings Plans rates.
}
\section{Request syntax}{
\preformatted{svc$describe_savings_plan_rates(
  savingsPlanId = "string",
  filters = list(
    list(
      name = "region"|"instanceType"|"productDescription"|"tenancy"|"productType"|"serviceCode"|"usageType"|"operation",
      values = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
