% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_describe_savings_plans}
\alias{savingsplans_describe_savings_plans}
\title{Describes the specified Savings Plans}
\usage{
savingsplans_describe_savings_plans(savingsPlanArns, savingsPlanIds,
  nextToken, maxResults, states, filters)
}
\arguments{
\item{savingsPlanArns}{The Amazon Resource Names (ARN) of the Savings Plans.}

\item{savingsPlanIds}{The IDs of the Savings Plans.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
additional results, make another call with the returned token value.}

\item{states}{The states.}

\item{filters}{The filters.}
}
\description{
Describes the specified Savings Plans.
}
\section{Request syntax}{
\preformatted{svc$describe_savings_plans(
  savingsPlanArns = list(
    "string"
  ),
  savingsPlanIds = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123,
  states = list(
    "payment-pending"|"payment-failed"|"active"|"retired"|"queued"|"queued-deleted"
  ),
  filters = list(
    list(
      name = "region"|"ec2-instance-family"|"commitment"|"upfront"|"term"|"savings-plan-type"|"payment-option"|"start"|"end",
      values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
