% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_describe_savings_plans_offering_rates}
\alias{savingsplans_describe_savings_plans_offering_rates}
\title{Describes the specified Savings Plans offering rates}
\usage{
savingsplans_describe_savings_plans_offering_rates(
  savingsPlanOfferingIds, savingsPlanPaymentOptions, savingsPlanTypes,
  products, serviceCodes, usageTypes, operations, filters, nextToken,
  maxResults)
}
\arguments{
\item{savingsPlanOfferingIds}{The IDs of the offerings.}

\item{savingsPlanPaymentOptions}{The payment options.}

\item{savingsPlanTypes}{The plan types.}

\item{products}{The AWS products.}

\item{serviceCodes}{The services.}

\item{usageTypes}{The usage details of the line item in the billing report.}

\item{operations}{The specific AWS operation for the line item in the billing report.}

\item{filters}{The filters.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
additional results, make another call with the returned token value.}
}
\description{
Describes the specified Savings Plans offering rates.
}
\section{Request syntax}{
\preformatted{svc$describe_savings_plans_offering_rates(
  savingsPlanOfferingIds = list(
    "string"
  ),
  savingsPlanPaymentOptions = list(
    "All Upfront"|"Partial Upfront"|"No Upfront"
  ),
  savingsPlanTypes = list(
    "Compute"|"EC2Instance"
  ),
  products = list(
    "EC2"|"Fargate"|"Lambda"
  ),
  serviceCodes = list(
    "AmazonEC2"|"AmazonECS"|"AWSLambda"
  ),
  usageTypes = list(
    "string"
  ),
  operations = list(
    "string"
  ),
  filters = list(
    list(
      name = "region"|"instanceFamily"|"instanceType"|"productDescription"|"tenancy"|"productId",
      values = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
