% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_describe_savings_plans_offerings}
\alias{savingsplans_describe_savings_plans_offerings}
\title{Describes the specified Savings Plans offerings}
\usage{
savingsplans_describe_savings_plans_offerings(offeringIds,
  paymentOptions, productType, planTypes, durations, currencies,
  descriptions, serviceCodes, usageTypes, operations, filters, nextToken,
  maxResults)
}
\arguments{
\item{offeringIds}{The IDs of the offerings.}

\item{paymentOptions}{The payment options.}

\item{productType}{The product type.}

\item{planTypes}{The plan type.}

\item{durations}{The durations, in seconds.}

\item{currencies}{The currencies.}

\item{descriptions}{The descriptions.}

\item{serviceCodes}{The services.}

\item{usageTypes}{The usage details of the line item in the billing report.}

\item{operations}{The specific AWS operation for the line item in the billing report.}

\item{filters}{The filters.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
additional results, make another call with the returned token value.}
}
\description{
Describes the specified Savings Plans offerings.
}
\section{Request syntax}{
\preformatted{svc$describe_savings_plans_offerings(
  offeringIds = list(
    "string"
  ),
  paymentOptions = list(
    "All Upfront"|"Partial Upfront"|"No Upfront"
  ),
  productType = "EC2"|"Fargate"|"Lambda",
  planTypes = list(
    "Compute"|"EC2Instance"
  ),
  durations = list(
    123
  ),
  currencies = list(
    "CNY"|"USD"
  ),
  descriptions = list(
    "string"
  ),
  serviceCodes = list(
    "string"
  ),
  usageTypes = list(
    "string"
  ),
  operations = list(
    "string"
  ),
  filters = list(
    list(
      name = "region"|"instanceFamily",
      values = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
