% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_untag_resource}
\alias{savingsplans_untag_resource}
\title{Removes the specified tags from the specified resource}
\usage{
savingsplans_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tagKeys}{[required] The tag keys.}
}
\description{
Removes the specified tags from the specified resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
