% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_create_discoverer}
\alias{schemas_create_discoverer}
\title{Creates a discoverer}
\usage{
schemas_create_discoverer(Description, SourceArn, Tags)
}
\arguments{
\item{Description}{A description for the discoverer.}

\item{SourceArn}{[required] The ARN of the event bus.}

\item{Tags}{Tags associated with the resource.}
}
\description{
Creates a discoverer.
}
\section{Request syntax}{
\preformatted{svc$create_discoverer(
  Description = "string",
  SourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
