% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_create_registry}
\alias{schemas_create_registry}
\title{Creates a registry}
\usage{
schemas_create_registry(Description, RegistryName, Tags)
}
\arguments{
\item{Description}{A description of the registry to be created.}

\item{RegistryName}{[required]}

\item{Tags}{Tags to associate with the registry.}
}
\description{
Creates a registry.
}
\section{Request syntax}{
\preformatted{svc$create_registry(
  Description = "string",
  RegistryName = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
