% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_delete_discoverer}
\alias{schemas_delete_discoverer}
\title{Deletes a discoverer}
\usage{
schemas_delete_discoverer(DiscovererId)
}
\arguments{
\item{DiscovererId}{[required]}
}
\description{
Deletes a discoverer.
}
\section{Request syntax}{
\preformatted{svc$delete_discoverer(
  DiscovererId = "string"
)
}
}

\keyword{internal}
