% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_delete_registry}
\alias{schemas_delete_registry}
\title{Deletes a Registry}
\usage{
schemas_delete_registry(RegistryName)
}
\arguments{
\item{RegistryName}{[required] The name of the registry.}
}
\description{
Deletes a Registry.
}
\section{Request syntax}{
\preformatted{svc$delete_registry(
  RegistryName = "string"
)
}
}

\keyword{internal}
