% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_delete_resource_policy}
\alias{schemas_delete_resource_policy}
\title{Delete the resource-based policy attached to the specified registry}
\usage{
schemas_delete_resource_policy(RegistryName)
}
\arguments{
\item{RegistryName}{The name of the registry.}
}
\description{
Delete the resource-based policy attached to the specified registry.
}
\section{Request syntax}{
\preformatted{svc$delete_resource_policy(
  RegistryName = "string"
)
}
}

\keyword{internal}
