% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_describe_discoverer}
\alias{schemas_describe_discoverer}
\title{Describes the discoverer}
\usage{
schemas_describe_discoverer(DiscovererId)
}
\arguments{
\item{DiscovererId}{[required]}
}
\description{
Describes the discoverer.
}
\section{Request syntax}{
\preformatted{svc$describe_discoverer(
  DiscovererId = "string"
)
}
}

\keyword{internal}
