% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_describe_schema}
\alias{schemas_describe_schema}
\title{Retrieve the schema definition}
\usage{
schemas_describe_schema(RegistryName, SchemaName, SchemaVersion)
}
\arguments{
\item{RegistryName}{[required]}

\item{SchemaName}{[required]}

\item{SchemaVersion}{}
}
\description{
Retrieve the schema definition.
}
\section{Request syntax}{
\preformatted{svc$describe_schema(
  RegistryName = "string",
  SchemaName = "string",
  SchemaVersion = "string"
)
}
}

\keyword{internal}
