% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_export_schema}
\alias{schemas_export_schema}
\title{Export schema}
\usage{
schemas_export_schema(RegistryName, SchemaName, SchemaVersion, Type)
}
\arguments{
\item{RegistryName}{[required] The name of the registry.}

\item{SchemaName}{[required] The name of the schema.}

\item{SchemaVersion}{Specifying this limits the results to only this schema version.}

\item{Type}{[required]}
}
\description{
Export schema
}
\section{Request syntax}{
\preformatted{svc$export_schema(
  RegistryName = "string",
  SchemaName = "string",
  SchemaVersion = "string",
  Type = "string"
)
}
}

\keyword{internal}
