% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_get_discovered_schema}
\alias{schemas_get_discovered_schema}
\title{Get the discovered schema that was generated based on sampled events}
\usage{
schemas_get_discovered_schema(Events, Type)
}
\arguments{
\item{Events}{[required] An array of strings that}

\item{Type}{[required] The type of event.}
}
\description{
Get the discovered schema that was generated based on sampled events.
}
\section{Request syntax}{
\preformatted{svc$get_discovered_schema(
  Events = list(
    "string"
  ),
  Type = "OpenApi3"
)
}
}

\keyword{internal}
