% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_list_registries}
\alias{schemas_list_registries}
\title{List the registries}
\usage{
schemas_list_registries(Limit, NextToken, RegistryNamePrefix, Scope)
}
\arguments{
\item{Limit}{}

\item{NextToken}{}

\item{RegistryNamePrefix}{}

\item{Scope}{}
}
\description{
List the registries.
}
\section{Request syntax}{
\preformatted{svc$list_registries(
  Limit = 123,
  NextToken = "string",
  RegistryNamePrefix = "string",
  Scope = "string"
)
}
}

\keyword{internal}
