% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_list_schema_versions}
\alias{schemas_list_schema_versions}
\title{Provides a list of the schema versions and related information}
\usage{
schemas_list_schema_versions(Limit, NextToken, RegistryName, SchemaName)
}
\arguments{
\item{Limit}{}

\item{NextToken}{}

\item{RegistryName}{[required]}

\item{SchemaName}{[required]}
}
\description{
Provides a list of the schema versions and related information.
}
\section{Request syntax}{
\preformatted{svc$list_schema_versions(
  Limit = 123,
  NextToken = "string",
  RegistryName = "string",
  SchemaName = "string"
)
}
}

\keyword{internal}
