% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_list_tags_for_resource}
\alias{schemas_list_tags_for_resource}
\title{Get tags for resource}
\usage{
schemas_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required]}
}
\description{
Get tags for resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
