% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_lock_service_linked_role}
\alias{schemas_lock_service_linked_role}
\title{Lock service linked role}
\usage{
schemas_lock_service_linked_role(RoleArn, Timeout)
}
\arguments{
\item{RoleArn}{[required]}

\item{Timeout}{[required]}
}
\description{
Lock service linked role
}
\section{Request syntax}{
\preformatted{svc$lock_service_linked_role(
  RoleArn = "string",
  Timeout = 123
)
}
}

\keyword{internal}
