% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_put_code_binding}
\alias{schemas_put_code_binding}
\title{Put code binding URI}
\usage{
schemas_put_code_binding(Language, RegistryName, SchemaName,
  SchemaVersion)
}
\arguments{
\item{Language}{[required]}

\item{RegistryName}{[required]}

\item{SchemaName}{[required]}

\item{SchemaVersion}{}
}
\description{
Put code binding URI
}
\section{Request syntax}{
\preformatted{svc$put_code_binding(
  Language = "string",
  RegistryName = "string",
  SchemaName = "string",
  SchemaVersion = "string"
)
}
}

\keyword{internal}
