% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_put_resource_policy}
\alias{schemas_put_resource_policy}
\title{The name of the policy}
\usage{
schemas_put_resource_policy(Policy, RegistryName, RevisionId)
}
\arguments{
\item{Policy}{[required] The resource-based policy.}

\item{RegistryName}{The name of the registry.}

\item{RevisionId}{The revision ID of the policy.}
}
\description{
The name of the policy.
}
\section{Request syntax}{
\preformatted{svc$put_resource_policy(
  Policy = "string",
  RegistryName = "string",
  RevisionId = "string"
)
}
}

\keyword{internal}
