% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_search_schemas}
\alias{schemas_search_schemas}
\title{Search the schemas}
\usage{
schemas_search_schemas(Keywords, Limit, NextToken, RegistryName)
}
\arguments{
\item{Keywords}{[required]}

\item{Limit}{}

\item{NextToken}{}

\item{RegistryName}{[required]}
}
\description{
Search the schemas
}
\section{Request syntax}{
\preformatted{svc$search_schemas(
  Keywords = "string",
  Limit = 123,
  NextToken = "string",
  RegistryName = "string"
)
}
}

\keyword{internal}
