% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_start_discoverer}
\alias{schemas_start_discoverer}
\title{Starts the discoverer}
\usage{
schemas_start_discoverer(DiscovererId)
}
\arguments{
\item{DiscovererId}{[required]}
}
\description{
Starts the discoverer
}
\section{Request syntax}{
\preformatted{svc$start_discoverer(
  DiscovererId = "string"
)
}
}

\keyword{internal}
