% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_tag_resource}
\alias{schemas_tag_resource}
\title{Add tags to a resource}
\usage{
schemas_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource.}

\item{Tags}{[required] Tags associated with the resource.}
}
\description{
Add tags to a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
