% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_unlock_service_linked_role}
\alias{schemas_unlock_service_linked_role}
\title{Unlock service linked role}
\usage{
schemas_unlock_service_linked_role(RoleArn)
}
\arguments{
\item{RoleArn}{[required]}
}
\description{
Unlock service linked role
}
\section{Request syntax}{
\preformatted{svc$unlock_service_linked_role(
  RoleArn = "string"
)
}
}

\keyword{internal}
