% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_untag_resource}
\alias{schemas_untag_resource}
\title{Removes tags from a resource}
\usage{
schemas_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required]}

\item{TagKeys}{[required]}
}
\description{
Removes tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
