% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_update_discoverer}
\alias{schemas_update_discoverer}
\title{Updates the discoverer}
\usage{
schemas_update_discoverer(Description, DiscovererId)
}
\arguments{
\item{Description}{The description of the discoverer to update.}

\item{DiscovererId}{[required]}
}
\description{
Updates the discoverer
}
\section{Request syntax}{
\preformatted{svc$update_discoverer(
  Description = "string",
  DiscovererId = "string"
)
}
}

\keyword{internal}
