% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_update_schema}
\alias{schemas_update_schema}
\title{Updates the schema definition}
\usage{
schemas_update_schema(ClientTokenId, Content, Description, RegistryName,
  SchemaName, Type)
}
\arguments{
\item{ClientTokenId}{The ID of the client token.}

\item{Content}{The source of the schema definition.}

\item{Description}{The description of the schema.}

\item{RegistryName}{[required]}

\item{SchemaName}{[required]}

\item{Type}{The schema type for the events schema.}
}
\description{
Updates the schema definition
}
\section{Request syntax}{
\preformatted{svc$update_schema(
  ClientTokenId = "string",
  Content = "string",
  Description = "string",
  RegistryName = "string",
  SchemaName = "string",
  Type = "OpenApi3"
)
}
}

\keyword{internal}
