% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_enable_standards}
\alias{securityhub_batch_enable_standards}
\title{Enables the standards specified by the standards ARNs}
\usage{
securityhub_batch_enable_standards(StandardsSubscriptionRequests)
}
\arguments{
\item{StandardsSubscriptionRequests}{[required] The list of standards that you want to enable.}
}
\description{
Enables the standards specified by the standards ARNs. In the context of
Security Hub, supported standards (for example, CIS AWS Foundations) are
automated and continuous checks that help determine your compliance
status against security industry (including AWS) best practices.
}
\section{Request syntax}{
\preformatted{securityhub$batch_enable_standards(
  StandardsSubscriptionRequests = list(
    list(
      StandardsArn = "string",
      StandardsInput = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
