% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_create_action_target}
\alias{securityhub_create_action_target}
\title{Creates a custom action target in Security Hub}
\usage{
securityhub_create_action_target(Name, Description, Id)
}
\arguments{
\item{Name}{[required] The name of the custom action target.}

\item{Description}{[required] The description for the custom action target.}

\item{Id}{[required] The ID for the custom action target.}
}
\description{
Creates a custom action target in Security Hub.
}
\details{
You can use custom actions on findings and insights in Security Hub to
trigger target actions in Amazon CloudWatch Events.
}
\section{Request syntax}{
\preformatted{svc$create_action_target(
  Name = "string",
  Description = "string",
  Id = "string"
)
}
}

\keyword{internal}
