% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_decline_invitations}
\alias{securityhub_decline_invitations}
\title{Declines invitations that are sent to this AWS account (invitee) from
the AWS accounts (inviters) that are specified by the provided
AccountIds}
\usage{
securityhub_decline_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{A list of account IDs that specify the accounts from which invitations
to Security Hub are declined.}
}
\description{
Declines invitations that are sent to this AWS account (invitee) from
the AWS accounts (inviters) that are specified by the provided
\code{AccountIds}.
}
\section{Request syntax}{
\preformatted{svc$decline_invitations(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
