% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_invitations}
\alias{securityhub_delete_invitations}
\title{Deletes invitations received by the AWS account to become a member
account}
\usage{
securityhub_delete_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{A list of the account IDs that sent the invitations to delete.}
}
\description{
Deletes invitations received by the AWS account to become a member
account.
}
\section{Request syntax}{
\preformatted{svc$delete_invitations(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
