% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_hub}
\alias{securityhub_describe_hub}
\title{Returns details about the Hub resource in your account, including the
HubArn and the time when you enabled Security Hub}
\usage{
securityhub_describe_hub(HubArn)
}
\arguments{
\item{HubArn}{The ARN of the Hub resource to retrieve.}
}
\description{
Returns details about the Hub resource in your account, including the
\code{HubArn} and the time when you enabled Security Hub.
}
\section{Request syntax}{
\preformatted{svc$describe_hub(
  HubArn = "string"
)
}
}

\keyword{internal}
