% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_products}
\alias{securityhub_describe_products}
\title{Returns information about the products available that you can subscribe
to}
\usage{
securityhub_describe_products(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token that is required for pagination.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Returns information about the products available that you can subscribe
to.
}
\section{Request syntax}{
\preformatted{svc$describe_products(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
