% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disable_import_findings_for_product}
\alias{securityhub_disable_import_findings_for_product}
\title{Stops you from being able to import findings generated by integrated
third-party providers into Security Hub}
\usage{
securityhub_disable_import_findings_for_product(ProductSubscriptionArn)
}
\arguments{
\item{ProductSubscriptionArn}{[required] The ARN of a resource that represents your subscription to a supported
product.}
}
\description{
Stops you from being able to import findings generated by integrated
third-party providers into Security Hub.
}
\section{Request syntax}{
\preformatted{securityhub$disable_import_findings_for_product(
  ProductSubscriptionArn = "string"
)
}
}

\keyword{internal}
