% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disassociate_from_master_account}
\alias{securityhub_disassociate_from_master_account}
\title{Disassociates the current Security Hub member account from its master
account}
\usage{
securityhub_disassociate_from_master_account()
}
\description{
Disassociates the current Security Hub member account from its master
account.
}
\section{Request syntax}{
\preformatted{securityhub$disassociate_from_master_account()
}
}

\keyword{internal}
