% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_enable_security_hub}
\alias{securityhub_enable_security_hub}
\title{Enables Security Hub for your account in the current Region or the
Region you specify in the request}
\usage{
securityhub_enable_security_hub(Tags)
}
\arguments{
\item{Tags}{The tags to add to the Hub resource when you enable Security Hub.}
}
\description{
Enables Security Hub for your account in the current Region or the
Region you specify in the request. Enabling Security Hub also enables
the CIS AWS Foundations standard. When you enable Security Hub, you
grant to Security Hub the permissions necessary to gather findings from
AWS Config, Amazon GuardDuty, Amazon Inspector, and Amazon Macie. To
learn more, see \href{https://docs.aws.amazon.com/securityhub/latest/userguide/securityhub-settingup.html}{Setting Up AWS Security Hub}.
}
\section{Request syntax}{
\preformatted{svc$enable_security_hub(
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
