% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_insight_results}
\alias{securityhub_get_insight_results}
\title{Lists the results of the Security Hub insight specified by the insight
ARN}
\usage{
securityhub_get_insight_results(InsightArn)
}
\arguments{
\item{InsightArn}{[required] The ARN of the insight whose results you want to see.}
}
\description{
Lists the results of the Security Hub insight specified by the insight
ARN.
}
\section{Request syntax}{
\preformatted{svc$get_insight_results(
  InsightArn = "string"
)
}
}

\keyword{internal}
