% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_members}
\alias{securityhub_get_members}
\title{Returns the details on the Security Hub member accounts that the account
IDs specify}
\usage{
securityhub_get_members(AccountIds)
}
\arguments{
\item{AccountIds}{[required] A list of account IDs for the Security Hub member accounts that you want
to return the details for.}
}
\description{
Returns the details on the Security Hub member accounts that the account
IDs specify.
}
\section{Request syntax}{
\preformatted{svc$get_members(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
