% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_product_subscribers}
\alias{securityhub_list_product_subscribers}
\title{Returns a list of account IDs that are subscribed to the product}
\usage{
securityhub_list_product_subscribers(ProductArn, NextToken, MaxResults)
}
\arguments{
\item{ProductArn}{The ARN of the product.}

\item{NextToken}{The token that is required for pagination.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Returns a list of account IDs that are subscribed to the product.
}
\section{Request syntax}{
\preformatted{svc$list_product_subscribers(
  ProductArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
