% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_tag_resource}
\alias{securityhub_tag_resource}
\title{Adds one or more tags to a resource}
\usage{
securityhub_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource to apply the tags to.}

\item{Tags}{[required] The tags to add to the resource.}
}
\description{
Adds one or more tags to a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
