% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_list_applications}
\alias{serverlessapplicationrepository_list_applications}
\title{Lists applications owned by the requester}
\usage{
serverlessapplicationrepository_list_applications(MaxItems, NextToken)
}
\arguments{
\item{MaxItems}{The total number of items to return.}

\item{NextToken}{A token to specify where to start paginating.}
}
\description{
Lists applications owned by the requester.
}
\section{Request syntax}{
\preformatted{serverlessapplicationrepository$list_applications(
  MaxItems = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
