% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_service_action}
\alias{servicecatalog_describe_service_action}
\title{Describes a self-service action}
\usage{
servicecatalog_describe_service_action(Id, AcceptLanguage)
}
\arguments{
\item{Id}{[required] The self-service action identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Describes a self-service action.
}
\section{Request syntax}{
\preformatted{servicecatalog$describe_service_action(
  Id = "string",
  AcceptLanguage = "string"
)
}
}

\keyword{internal}
