% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_get_aws_organizations_access_status}
\alias{servicecatalog_get_aws_organizations_access_status}
\title{Get the Access Status for AWS Organization portfolio share feature}
\usage{
servicecatalog_get_aws_organizations_access_status()
}
\description{
Get the Access Status for AWS Organization portfolio share feature. This
API can only be called by the master account in the organization or by a
delegated admin.
}
\section{Request syntax}{
\preformatted{svc$get_aws_organizations_access_status()
}
}

\keyword{internal}
