% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_update_service_action}
\alias{servicecatalog_update_service_action}
\title{Updates a self-service action}
\usage{
servicecatalog_update_service_action(Id, Name, Definition, Description,
  AcceptLanguage)
}
\arguments{
\item{Id}{[required] The self-service action identifier.}

\item{Name}{The self-service action name.}

\item{Definition}{A map that defines the self-service action.}

\item{Description}{The self-service action description.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Updates a self-service action.
}
\section{Request syntax}{
\preformatted{svc$update_service_action(
  Id = "string",
  Name = "string",
  Definition = list(
    "string"
  ),
  Description = "string",
  AcceptLanguage = "string"
)
}
}

\keyword{internal}
