% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_get_namespace}
\alias{servicediscovery_get_namespace}
\title{Gets information about a namespace}
\usage{
servicediscovery_get_namespace(Id)
}
\arguments{
\item{Id}{[required] The ID of the namespace that you want to get information about.}
}
\description{
Gets information about a namespace.
}
\section{Request syntax}{
\preformatted{svc$get_namespace(
  Id = "string"
)
}
}

\examples{
\dontrun{
# This example gets information about a specified namespace.
svc$get_namespace(
  Id = "ns-e4anhexample0004"
)
}

}
\keyword{internal}
