% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_requested_service_quota_change}
\alias{servicequotas_get_requested_service_quota_change}
\title{Retrieves the details for a particular increase request}
\usage{
servicequotas_get_requested_service_quota_change(RequestId)
}
\arguments{
\item{RequestId}{[required] Identifies the quota increase request.}
}
\description{
Retrieves the details for a particular increase request.
}
\section{Request syntax}{
\preformatted{svc$get_requested_service_quota_change(
  RequestId = "string"
)
}
}

\keyword{internal}
