% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_service_quota_increase_request_from_template}
\alias{servicequotas_get_service_quota_increase_request_from_template}
\title{Retrieves information about the specified quota increase request in your
quota request template}
\usage{
servicequotas_get_service_quota_increase_request_from_template(
  ServiceCode, QuotaCode, AwsRegion)
}
\arguments{
\item{ServiceCode}{[required] The service identifier.}

\item{QuotaCode}{[required] The quota identifier.}

\item{AwsRegion}{[required] The AWS Region.}
}
\description{
Retrieves information about the specified quota increase request in your
quota request template.
}
\section{Request syntax}{
\preformatted{svc$get_service_quota_increase_request_from_template(
  ServiceCode = "string",
  QuotaCode = "string",
  AwsRegion = "string"
)
}
}

\keyword{internal}
