% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_request_service_quota_increase}
\alias{servicequotas_request_service_quota_increase}
\title{Retrieves the details of a service quota increase request}
\usage{
servicequotas_request_service_quota_increase(ServiceCode, QuotaCode,
  DesiredValue)
}
\arguments{
\item{ServiceCode}{[required] Specifies the service that you want to use.}

\item{QuotaCode}{[required] Specifies the service quota that you want to use.}

\item{DesiredValue}{[required] Specifies the value submitted in the service quota increase request.}
}
\description{
Retrieves the details of a service quota increase request. The response
to this command provides the details in the RequestedServiceQuotaChange
object.
}
\section{Request syntax}{
\preformatted{svc$request_service_quota_increase(
  ServiceCode = "string",
  QuotaCode = "string",
  DesiredValue = 123.0
)
}
}

\keyword{internal}
