% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_identity}
\alias{ses_delete_identity}
\title{Deletes the specified identity (an email address or a domain) from the
list of verified identities}
\usage{
ses_delete_identity(Identity)
}
\arguments{
\item{Identity}{[required] The identity to be removed from the list of identities for the AWS
Account.}
}
\description{
Deletes the specified identity (an email address or a domain) from the
list of verified identities.

You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$delete_identity(
  Identity = "string"
)
}
}

\examples{
\dontrun{
# The following example deletes an identity from the list of identities
# that have been submitted for verification with Amazon SES:
svc$delete_identity(
  Identity = "user@example.com"
)
}

}
\keyword{internal}
