% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_template}
\alias{ses_delete_template}
\title{Deletes an email template}
\usage{
ses_delete_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the template to be deleted.}
}
\description{
Deletes an email template.
}
\details{
You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{ses$delete_template(
  TemplateName = "string"
)
}
}

\keyword{internal}
