% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_template}
\alias{ses_get_template}
\title{Displays the template object (which includes the Subject line, HTML part
and text part) for the template you specify}
\usage{
ses_get_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the template you want to retrieve.}
}
\description{
Displays the template object (which includes the Subject line, HTML part
and text part) for the template you specify.
}
\details{
You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{ses$get_template(
  TemplateName = "string"
)
}
}

\keyword{internal}
