% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_update_account_sending_enabled}
\alias{ses_update_account_sending_enabled}
\title{Enables or disables email sending across your entire Amazon SES account
in the current AWS Region}
\usage{
ses_update_account_sending_enabled(Enabled)
}
\arguments{
\item{Enabled}{Describes whether email sending is enabled or disabled for your Amazon
SES account in the current AWS Region.}
}
\description{
Enables or disables email sending across your entire Amazon SES account
in the current AWS Region. You can use this operation in conjunction
with Amazon CloudWatch alarms to temporarily pause email sending across
your Amazon SES account in a given AWS Region when reputation metrics
(such as your bounce or complaint rates) reach certain thresholds.
}
\details{
You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$update_account_sending_enabled(
  Enabled = TRUE|FALSE
)
}
}

\examples{
# The following example updated the sending status for this account.
\donttest{svc$update_account_sending_enabled(
  Enabled = TRUE
)}

}
\keyword{internal}
