% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_verify_email_address}
\alias{ses_verify_email_address}
\title{Deprecated}
\usage{
ses_verify_email_address(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] The email address to be verified.}
}
\description{
Deprecated. Use the \code{VerifyEmailIdentity} operation to verify a new
email address.
}
\section{Request syntax}{
\preformatted{svc$verify_email_address(
  EmailAddress = "string"
)
}
}

\section{Examples}{
\preformatted{# The following example starts the email address verification process with
# Amazon SES:
svc$verify_email_address(
  EmailAddress = "user@example.com"
)
}
}

\keyword{internal}
