% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_contact_list}
\alias{sesv2_create_contact_list}
\title{Creates a contact list}
\usage{
sesv2_create_contact_list(ContactListName, Topics, Description, Tags)
}
\arguments{
\item{ContactListName}{[required] The name of the contact list.}

\item{Topics}{An interest group, theme, or label within a list. A contact list can
have multiple topics.}

\item{Description}{A description of what the contact list is about.}

\item{Tags}{The tags associated with a contact list.}
}
\description{
Creates a contact list.
}
\section{Request syntax}{
\preformatted{svc$create_contact_list(
  ContactListName = "string",
  Topics = list(
    list(
      TopicName = "string",
      DisplayName = "string",
      Description = "string",
      DefaultSubscriptionStatus = "OPT_IN"|"OPT_OUT"
    )
  ),
  Description = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
