% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_email_identity}
\alias{sesv2_create_email_identity}
\title{Starts the process of verifying an email identity}
\usage{
sesv2_create_email_identity(EmailIdentity, Tags)
}
\arguments{
\item{EmailIdentity}{[required] The email address or domain that you want to verify.}

\item{Tags}{An array of objects that define the tags (keys and values) that you want
to associate with the email identity.}
}
\description{
Starts the process of verifying an email identity. An \emph{identity} is an
email address or domain that you use when you send email. Before you can
use an identity to send email, you first have to verify it. By verifying
an identity, you demonstrate that you\'re the owner of the identity, and
that you\'ve given Amazon SES API v2 permission to send email from the
identity.
}
\details{
When you verify an email address, Amazon SES sends an email to the
address. Your email address is verified as soon as you follow the link
in the verification email.

When you verify a domain, this operation provides a set of DKIM tokens,
which you can convert into CNAME tokens. You add these CNAME tokens to
the DNS configuration for your domain. Your domain is verified when
Amazon SES detects these records in the DNS configuration for your
domain. For some DNS providers, it can take 72 hours or more to complete
the domain verification process.
}
\section{Request syntax}{
\preformatted{svc$create_email_identity(
  EmailIdentity = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
