% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_email_template}
\alias{sesv2_create_email_template}
\title{Creates an email template}
\usage{
sesv2_create_email_template(TemplateName, TemplateContent)
}
\arguments{
\item{TemplateName}{[required] The name of the template you want to create.}

\item{TemplateContent}{[required] The content of the email template, composed of a subject line, an HTML
part, and a text-only part.}
}
\description{
Creates an email template. Email templates enable you to send
personalized email to one or more destinations in a single API
operation. For more information, see the \href{https://docs.aws.amazon.com/ses/latest/DeveloperGuide/send-personalized-email-api.html}{Amazon SES Developer Guide}.
}
\details{
You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$create_email_template(
  TemplateName = "string",
  TemplateContent = list(
    Subject = "string",
    Text = "string",
    Html = "string"
  )
)
}
}

\keyword{internal}
