% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_configuration_set}
\alias{sesv2_delete_configuration_set}
\title{Delete an existing configuration set}
\usage{
sesv2_delete_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to delete.}
}
\description{
Delete an existing configuration set.
}
\details{
\emph{Configuration sets} are groups of rules that you can apply to the
emails you send. You apply a configuration set to an email by including
a reference to the configuration set in the headers of the email. When
you apply a configuration set to an email, all of the rules in that
configuration set are applied to the email.
}
\section{Request syntax}{
\preformatted{svc$delete_configuration_set(
  ConfigurationSetName = "string"
)
}
}

\keyword{internal}
