% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_contact}
\alias{sesv2_delete_contact}
\title{Removes a contact from a contact list}
\usage{
sesv2_delete_contact(ContactListName, EmailAddress)
}
\arguments{
\item{ContactListName}{[required] The name of the contact list from which the contact should be removed.}

\item{EmailAddress}{[required] The contact's email address.}
}
\description{
Removes a contact from a contact list.
}
\section{Request syntax}{
\preformatted{svc$delete_contact(
  ContactListName = "string",
  EmailAddress = "string"
)
}
}

\keyword{internal}
