% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_custom_verification_email_template}
\alias{sesv2_delete_custom_verification_email_template}
\title{Deletes an existing custom verification email template}
\usage{
sesv2_delete_custom_verification_email_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the custom verification email template that you want to
delete.}
}
\description{
Deletes an existing custom verification email template.

For more information about custom verification email templates, see
Using Custom Verification Email Templates in the \emph{Amazon SES Developer
Guide}.

You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$delete_custom_verification_email_template(
  TemplateName = "string"
)
}
}

\keyword{internal}
