% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_dedicated_ip_pool}
\alias{sesv2_delete_dedicated_ip_pool}
\title{Delete a dedicated IP pool}
\usage{
sesv2_delete_dedicated_ip_pool(PoolName)
}
\arguments{
\item{PoolName}{[required] The name of the dedicated IP pool that you want to delete.}
}
\description{
Delete a dedicated IP pool.
}
\section{Request syntax}{
\preformatted{svc$delete_dedicated_ip_pool(
  PoolName = "string"
)
}
}

\keyword{internal}
