% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_email_identity_policy}
\alias{sesv2_delete_email_identity_policy}
\title{Deletes the specified sending authorization policy for the given
identity (an email address or a domain)}
\usage{
sesv2_delete_email_identity_policy(EmailIdentity, PolicyName)
}
\arguments{
\item{EmailIdentity}{[required] The email identity for which you want to delete a policy.}

\item{PolicyName}{[required] The name of the policy.

The policy name cannot exceed 64 characters and can only include
alphanumeric characters, dashes, and underscores.}
}
\description{
Deletes the specified sending authorization policy for the given
identity (an email address or a domain). This API returns successfully
even if a policy with the specified name does not exist.

This API is for the identity owner only. If you have not verified the
identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to
authorize other senders to use its identities. For information about
using sending authorization, see the \href{https://docs.aws.amazon.com/ses/latest/DeveloperGuide/sending-authorization.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$delete_email_identity_policy(
  EmailIdentity = "string",
  PolicyName = "string"
)
}
}

\keyword{internal}
