% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_suppressed_destination}
\alias{sesv2_delete_suppressed_destination}
\title{Used to delete a suppressed email destination from your suppression list}
\usage{
sesv2_delete_suppressed_destination(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] The suppressed email destination to delete.}
}
\description{
Used to delete a suppressed email destination from your suppression
list.
}
\section{Request syntax}{
\preformatted{svc$delete_suppressed_destination(
  EmailAddress = "string"
)
}
}

\keyword{internal}
