% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_account}
\alias{sesv2_get_account}
\title{Obtain information about the email-sending status and capabilities of
your Amazon SES account in the current AWS Region}
\usage{
sesv2_get_account()
}
\description{
Obtain information about the email-sending status and capabilities of
your Amazon SES account in the current AWS Region.
}
\section{Request syntax}{
\preformatted{svc$get_account()
}
}

\keyword{internal}
