% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_custom_verification_email_template}
\alias{sesv2_get_custom_verification_email_template}
\title{Returns the custom email verification template for the template name you
specify}
\usage{
sesv2_get_custom_verification_email_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the custom verification email template that you want to
retrieve.}
}
\description{
Returns the custom email verification template for the template name you
specify.
}
\details{
For more information about custom verification email templates, see
\href{https://docs.aws.amazon.com/ses/latest/DeveloperGuide/send-email-verify-address-custom.html}{Using Custom Verification Email Templates}
in the \emph{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$get_custom_verification_email_template(
  TemplateName = "string"
)
}
}

\keyword{internal}
