% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_dedicated_ips}
\alias{sesv2_get_dedicated_ips}
\title{List the dedicated IP addresses that are associated with your AWS
account}
\usage{
sesv2_get_dedicated_ips(PoolName, NextToken, PageSize)
}
\arguments{
\item{PoolName}{The name of the IP pool that the dedicated IP address is associated
with.}

\item{NextToken}{A token returned from a previous call to \code{GetDedicatedIps} to indicate
the position of the dedicated IP pool in the list of IP pools.}

\item{PageSize}{The number of results to show in a single call to
\code{GetDedicatedIpsRequest}. If the number of results is larger than the
number you specified in this parameter, then the response includes a
\code{NextToken} element, which you can use to obtain additional results.}
}
\description{
List the dedicated IP addresses that are associated with your AWS
account.
}
\section{Request syntax}{
\preformatted{svc$get_dedicated_ips(
  PoolName = "string",
  NextToken = "string",
  PageSize = 123
)
}
}

\keyword{internal}
