% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_deliverability_dashboard_options}
\alias{sesv2_get_deliverability_dashboard_options}
\title{Retrieve information about the status of the Deliverability dashboard
for your account}
\usage{
sesv2_get_deliverability_dashboard_options()
}
\description{
Retrieve information about the status of the Deliverability dashboard
for your account. When the Deliverability dashboard is enabled, you gain
access to reputation, deliverability, and other metrics for the domains
that you use to send email. You also gain the ability to perform
predictive inbox placement tests.
}
\details{
When you use the Deliverability dashboard, you pay a monthly
subscription charge, in addition to any other fees that you accrue by
using Amazon SES and other AWS services. For more information about the
features and cost of a Deliverability dashboard subscription, see
\href{https://aws.amazon.com/ses/pricing/}{Amazon SES Pricing}.
}
\section{Request syntax}{
\preformatted{svc$get_deliverability_dashboard_options()
}
}

\keyword{internal}
