% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_domain_deliverability_campaign}
\alias{sesv2_get_domain_deliverability_campaign}
\title{Retrieve all the deliverability data for a specific campaign}
\usage{
sesv2_get_domain_deliverability_campaign(CampaignId)
}
\arguments{
\item{CampaignId}{[required] The unique identifier for the campaign. The Deliverability dashboard
automatically generates and assigns this identifier to a campaign.}
}
\description{
Retrieve all the deliverability data for a specific campaign. This data
is available for a campaign only if the campaign sent email by using a
domain that the Deliverability dashboard is enabled for.
}
\section{Request syntax}{
\preformatted{svc$get_domain_deliverability_campaign(
  CampaignId = "string"
)
}
}

\keyword{internal}
