% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_email_identity}
\alias{sesv2_get_email_identity}
\title{Provides information about a specific identity, including the identity's
verification status, its DKIM authentication status, and its custom
Mail-From settings}
\usage{
sesv2_get_email_identity(EmailIdentity)
}
\arguments{
\item{EmailIdentity}{[required] The email identity that you want to retrieve details for.}
}
\description{
Provides information about a specific identity, including the identity's
verification status, its DKIM authentication status, and its custom
Mail-From settings.
}
\section{Request syntax}{
\preformatted{svc$get_email_identity(
  EmailIdentity = "string"
)
}
}

\keyword{internal}
