% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_email_identity_policies}
\alias{sesv2_get_email_identity_policies}
\title{Returns the requested sending authorization policies for the given
identity (an email address or a domain)}
\usage{
sesv2_get_email_identity_policies(EmailIdentity)
}
\arguments{
\item{EmailIdentity}{[required] The email identity that you want to retrieve policies for.}
}
\description{
Returns the requested sending authorization policies for the given
identity (an email address or a domain). The policies are returned as a
map of policy names to policy contents. You can retrieve a maximum of 20
policies at a time.

This API is for the identity owner only. If you have not verified the
identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to
authorize other senders to use its identities. For information about
using sending authorization, see the \href{https://docs.aws.amazon.com/ses/latest/DeveloperGuide/sending-authorization.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$get_email_identity_policies(
  EmailIdentity = "string"
)
}
}

\keyword{internal}
