% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_email_template}
\alias{sesv2_get_email_template}
\title{Displays the template object (which includes the subject line, HTML part
and text part) for the template you specify}
\usage{
sesv2_get_email_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the template you want to retrieve.}
}
\description{
Displays the template object (which includes the subject line, HTML part
and text part) for the template you specify.

You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$get_email_template(
  TemplateName = "string"
)
}
}

\keyword{internal}
