% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_suppressed_destination}
\alias{sesv2_get_suppressed_destination}
\title{Retrieves information about a specific email address that's on the
suppression list for your account}
\usage{
sesv2_get_suppressed_destination(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] The email address that\'s on the account suppression list.}
}
\description{
Retrieves information about a specific email address that\'s on the
suppression list for your account.
}
\section{Request syntax}{
\preformatted{svc$get_suppressed_destination(
  EmailAddress = "string"
)
}
}

\keyword{internal}
