% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_custom_verification_email_templates}
\alias{sesv2_list_custom_verification_email_templates}
\title{Lists the existing custom verification email templates for your account
in the current AWS Region}
\usage{
sesv2_list_custom_verification_email_templates(NextToken, PageSize)
}
\arguments{
\item{NextToken}{A token returned from a previous call to
\code{ListCustomVerificationEmailTemplates} to indicate the position in the
list of custom verification email templates.}

\item{PageSize}{The number of results to show in a single call to
\code{ListCustomVerificationEmailTemplates}. If the number of results is
larger than the number you specified in this parameter, then the
response includes a \code{NextToken} element, which you can use to obtain
additional results.

The value you specify has to be at least 1, and can be no more than 50.}
}
\description{
Lists the existing custom verification email templates for your account
in the current AWS Region.

For more information about custom verification email templates, see
\href{https://docs.aws.amazon.com/ses/latest/DeveloperGuide/send-email-verify-address-custom.html}{Using Custom Verification Email Templates}
in the \emph{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$list_custom_verification_email_templates(
  NextToken = "string",
  PageSize = 123
)
}
}

\keyword{internal}
