% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_dedicated_ip_pools}
\alias{sesv2_list_dedicated_ip_pools}
\title{List all of the dedicated IP pools that exist in your AWS account in the
current Region}
\usage{
sesv2_list_dedicated_ip_pools(NextToken, PageSize)
}
\arguments{
\item{NextToken}{A token returned from a previous call to \code{ListDedicatedIpPools} to
indicate the position in the list of dedicated IP pools.}

\item{PageSize}{The number of results to show in a single call to
\code{ListDedicatedIpPools}. If the number of results is larger than the
number you specified in this parameter, then the response includes a
\code{NextToken} element, which you can use to obtain additional results.}
}
\description{
List all of the dedicated IP pools that exist in your AWS account in the
current Region.
}
\section{Request syntax}{
\preformatted{svc$list_dedicated_ip_pools(
  NextToken = "string",
  PageSize = 123
)
}
}

\keyword{internal}
