% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_email_templates}
\alias{sesv2_list_email_templates}
\title{Lists the email templates present in your Amazon SES account in the
current AWS Region}
\usage{
sesv2_list_email_templates(NextToken, PageSize)
}
\arguments{
\item{NextToken}{A token returned from a previous call to \code{ListEmailTemplates} to
indicate the position in the list of email templates.}

\item{PageSize}{The number of results to show in a single call to \code{ListEmailTemplates}.
If the number of results is larger than the number you specified in this
parameter, then the response includes a \code{NextToken} element, which you
can use to obtain additional results.

The value you specify has to be at least 1, and can be no more than 10.}
}
\description{
Lists the email templates present in your Amazon SES account in the
current AWS Region.
}
\details{
You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$list_email_templates(
  NextToken = "string",
  PageSize = 123
)
}
}

\keyword{internal}
