% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_suppressed_destinations}
\alias{sesv2_list_suppressed_destinations}
\title{Used to fetch a list suppressed email destinations from your suppression
list}
\usage{
sesv2_list_suppressed_destinations(Reasons, StartDate, EndDate,
  NextToken, PageSize)
}
\arguments{
\item{Reasons}{Filters email destinations suppressed by the given reasons.}

\item{StartDate}{Filters email destinations suppressed before the given time.}

\item{EndDate}{Filters email destinations suppressed after the given time.}

\item{NextToken}{A token returned from a previous call to \code{ListSuppressedDestinations} to
indicate the position in the list of suppressed email destinations.}

\item{PageSize}{The number of results to show in a single call to
\code{ListSuppressedDestinations}. If the number of results is larger than
the number you specified in this parameter, then the response includes a
\code{NextToken} element, which you can use to obtain additional results.}
}
\description{
Used to fetch a list suppressed email destinations from your suppression
list.
}
\section{Request syntax}{
\preformatted{svc$list_suppressed_destinations(
  Reasons = list(
    "BOUNCE"|"COMPLAINT"
  ),
  StartDate = as.POSIXct(
    "2015-01-01"
  ),
  EndDate = as.POSIXct(
    "2015-01-01"
  ),
  NextToken = "string",
  PageSize = 123
)
}
}

\keyword{internal}
