% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_tags_for_resource}
\alias{sesv2_list_tags_for_resource}
\title{Retrieve a list of the tags (keys and values) that are associated with a
specified resource}
\usage{
sesv2_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to retrieve
tag information for.}
}
\description{
Retrieve a list of the tags (keys and values) that are associated with a
specified resource. A \emph{tag} is a label that you optionally define and
associate with a resource. Each tag consists of a required \emph{tag key} and
an optional associated \emph{tag value}. A tag key is a general label that
acts as a category for more specific tag values. A tag value acts as a
descriptor within a tag key.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
