% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_account_suppression_attributes}
\alias{sesv2_put_account_suppression_attributes}
\title{Change your account's suppression preferences for your account}
\usage{
sesv2_put_account_suppression_attributes(SuppressedReasons)
}
\arguments{
\item{SuppressedReasons}{A list of reasons to suppress email addresses. The only valid reasons
are:
\itemize{
\item \code{COMPLAINT} -- Amazon SES will suppress an email address that
receives a complaint.
\item \code{BOUNCE} -- Amazon SES will suppress an email address that hard
bounces.
}}
}
\description{
Change your account\'s suppression preferences for your account.
}
\section{Request syntax}{
\preformatted{svc$put_account_suppression_attributes(
  SuppressedReasons = list(
    "BOUNCE"|"COMPLAINT"
  )
)
}
}

\keyword{internal}
