% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_configuration_set_suppression_options}
\alias{sesv2_put_configuration_set_suppression_options}
\title{Specify your account's suppression preferences for a configuration set}
\usage{
sesv2_put_configuration_set_suppression_options(ConfigurationSetName,
  SuppressedReasons)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to enable or disable
email sending for.}

\item{SuppressedReasons}{A list of reasons to suppress email addresses. The only valid reasons
are:
\itemize{
\item \code{COMPLAINT} -- Amazon SES will suppress an email address that
receives a complaint.
\item \code{BOUNCE} -- Amazon SES will suppress an email address that hard
bounces.
}}
}
\description{
Specify your account\'s suppression preferences for a configuration set.
}
\section{Request syntax}{
\preformatted{svc$put_configuration_set_suppression_options(
  ConfigurationSetName = "string",
  SuppressedReasons = list(
    "BOUNCE"|"COMPLAINT"
  )
)
}
}

\keyword{internal}
