% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_email_identity_dkim_signing_attributes}
\alias{sesv2_put_email_identity_dkim_signing_attributes}
\title{Used to configure or change the DKIM authentication settings for an
email domain identity}
\usage{
sesv2_put_email_identity_dkim_signing_attributes(EmailIdentity,
  SigningAttributesOrigin, SigningAttributes)
}
\arguments{
\item{EmailIdentity}{[required] The email identity that you want to configure DKIM for.}

\item{SigningAttributesOrigin}{[required] The method that you want to use to configure DKIM for the identity.
There are two possible values:
\itemize{
\item \code{AWS_SES} – Configure DKIM for the identity by using \href{https://docs.aws.amazon.com/ses/latest/DeveloperGuide/send-email-authentication-dkim-easy.html}{Easy DKIM}.
\item \code{EXTERNAL} – Configure DKIM for the identity by using Bring Your Own
DKIM (BYODKIM).
}}

\item{SigningAttributes}{An object that contains information about the private key and selector
that you want to use to configure DKIM for the identity. This object is
only required if you want to configure Bring Your Own DKIM (BYODKIM) for
the identity.}
}
\description{
Used to configure or change the DKIM authentication settings for an
email domain identity. You can use this operation to do any of the
following:
}
\details{
\itemize{
\item Update the signing attributes for an identity that uses Bring Your
Own DKIM (BYODKIM).
\item Change from using no DKIM authentication to using Easy DKIM.
\item Change from using no DKIM authentication to using BYODKIM.
\item Change from using Easy DKIM to using BYODKIM.
\item Change from using BYODKIM to using Easy DKIM.
}
}
\section{Request syntax}{
\preformatted{svc$put_email_identity_dkim_signing_attributes(
  EmailIdentity = "string",
  SigningAttributesOrigin = "AWS_SES"|"EXTERNAL",
  SigningAttributes = list(
    DomainSigningSelector = "string",
    DomainSigningPrivateKey = "string"
  )
)
}
}

\keyword{internal}
