% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_suppressed_destination}
\alias{sesv2_put_suppressed_destination}
\title{Puts (overwrites) an email destination in your suppression list}
\usage{
sesv2_put_suppressed_destination(EmailAddress, Reason)
}
\arguments{
\item{EmailAddress}{[required] Email destination to be suppressed.}

\item{Reason}{[required] Reason for which the email destination is suppressed.}
}
\description{
Puts (overwrites) an email destination in your suppression list.
}
\section{Request syntax}{
\preformatted{svc$put_suppressed_destination(
  EmailAddress = "string",
  Reason = "BOUNCE"|"COMPLAINT"
)
}
}

\keyword{internal}
