% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_update_contact_list}
\alias{sesv2_update_contact_list}
\title{Updates contact list metadata}
\usage{
sesv2_update_contact_list(ContactListName, Topics, Description)
}
\arguments{
\item{ContactListName}{[required] The name of the contact list.}

\item{Topics}{An interest group, theme, or label within a list. A contact list can
have multiple topics.}

\item{Description}{A description of what the contact list is about.}
}
\description{
Updates contact list metadata. This operation does a complete
replacement.
}
\section{Request syntax}{
\preformatted{svc$update_contact_list(
  ContactListName = "string",
  Topics = list(
    list(
      TopicName = "string",
      DisplayName = "string",
      Description = "string",
      DefaultSubscriptionStatus = "OPT_IN"|"OPT_OUT"
    )
  ),
  Description = "string"
)
}
}

\keyword{internal}
