% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_parameters}
\alias{set_parameters}
\title{Convenience function for changing parameters}
\usage{
set_parameters(..., based.on = default_parameters())
}
\arguments{
\item{...}{name and value of parameters to set (e.g. fps=1500)}

\item{based.on}{an existing complete set of parameters to modify
(see \code{\link{default_parameters}}).}
}
\value{
the modified list of parameters
}
\description{
Convenience function for changing parameters
}
