% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_protection}
\alias{shield_describe_protection}
\title{Lists the details of a Protection object}
\usage{
shield_describe_protection(ProtectionId, ResourceArn)
}
\arguments{
\item{ProtectionId}{The unique identifier (ID) for the Protection object that is described.
When submitting the \code{DescribeProtection} request you must provide either
the \code{ResourceArn} or the \code{ProtectionID}, but not both.}

\item{ResourceArn}{The ARN (Amazon Resource Name) of the AWS resource for the Protection
object that is described. When submitting the \code{DescribeProtection}
request you must provide either the \code{ResourceArn} or the \code{ProtectionID},
but not both.}
}
\description{
Lists the details of a Protection object.
}
\section{Request syntax}{
\preformatted{shield$describe_protection(
  ProtectionId = "string",
  ResourceArn = "string"
)
}
}

\keyword{internal}
