% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_disable_proactive_engagement}
\alias{shield_disable_proactive_engagement}
\title{Removes authorization from the DDoS Response Team (DRT) to notify
contacts about escalations to the DRT and to initiate proactive customer
support}
\usage{
shield_disable_proactive_engagement()
}
\description{
Removes authorization from the DDoS Response Team (DRT) to notify
contacts about escalations to the DRT and to initiate proactive customer
support.
}
\section{Request syntax}{
\preformatted{svc$disable_proactive_engagement()
}
}

\keyword{internal}
