% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_disassociate_drt_role}
\alias{shield_disassociate_drt_role}
\title{Removes the DDoS Response Team's (DRT) access to your AWS account}
\usage{
shield_disassociate_drt_role()
}
\description{
Removes the DDoS Response Team's (DRT) access to your AWS account.
}
\details{
To make a \code{DisassociateDRTRole} request, you must be subscribed to the
\href{https://aws.amazon.com/premiumsupport/business-support/}{Business Support plan} or the
\href{https://aws.amazon.com/premiumsupport/enterprise-support/}{Enterprise Support plan}.
However, if you are not subscribed to one of these support plans, but
had been previously and had granted the DRT access to your account, you
can submit a \code{DisassociateDRTRole} request to remove this access.
}
\section{Request syntax}{
\preformatted{svc$disassociate_drt_role()
}
}

\keyword{internal}
