% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_get_signing_platform}
\alias{signer_get_signing_platform}
\title{Returns information on a specific signing platform}
\usage{
signer_get_signing_platform(platformId)
}
\arguments{
\item{platformId}{[required] The ID of the target signing platform.}
}
\description{
Returns information on a specific signing platform.
}
\section{Request syntax}{
\preformatted{signer$get_signing_platform(
  platformId = "string"
)
}
}

\keyword{internal}
