% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_start_signing_job}
\alias{signer_start_signing_job}
\title{Initiates a signing job to be performed on the code provided}
\usage{
signer_start_signing_job(source, destination, profileName,
  clientRequestToken)
}
\arguments{
\item{source}{[required] The S3 bucket that contains the object to sign or a BLOB that contains
your raw code.}

\item{destination}{[required] The S3 bucket in which to save your signed object. The destination
contains the name of your bucket and an optional prefix.}

\item{profileName}{The name of the signing profile.}

\item{clientRequestToken}{[required] String that identifies the signing request. All calls after the first
that use this token return the same response as the first call.}
}
\description{
Initiates a signing job to be performed on the code provided. Signing
jobs are viewable by the \code{ListSigningJobs} operation for two years after
they are performed. Note the following requirements:
}
\details{
\itemize{
\item You must create an Amazon S3 source bucket. For more information,
see \href{http://docs.aws.amazon.com/AmazonS3/latest/gsg/CreatingABucket.html}{Create a Bucket}
in the \emph{Amazon S3 Getting Started Guide}.
\item Your S3 source bucket must be version enabled.
\item You must create an S3 destination bucket. Code signing uses your S3
destination bucket to write your signed code.
\item You specify the name of the source and destination buckets when
calling the \code{StartSigningJob} operation.
\item You must also specify a request token that identifies your request
to code signing.
}

You can call the DescribeSigningJob and the ListSigningJobs actions
after you call \code{StartSigningJob}.

For a Java example that shows how to use this action, see
\url{http://docs.aws.amazon.com/acm/latest/userguide/}
}
\section{Request syntax}{
\preformatted{svc$start_signing_job(
  source = list(
    s3 = list(
      bucketName = "string",
      key = "string",
      version = "string"
    )
  ),
  destination = list(
    s3 = list(
      bucketName = "string",
      prefix = "string"
    )
  ),
  profileName = "string",
  clientRequestToken = "string"
)
}
}

\keyword{internal}
