% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_tag_resource}
\alias{signer_tag_resource}
\title{Adds one or more tags to a signing profile}
\usage{
signer_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] Amazon Resource Name (ARN) for the signing profile.}

\item{tags}{[required] One or more tags to be associated with the signing profile.}
}
\description{
Adds one or more tags to a signing profile. Tags are labels that you can
use to identify and organize your AWS resources. Each tag consists of a
key and an optional value. You specify the signing profile using its
Amazon Resource Name (ARN). You specify the tag by using a key-value
pair.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
