% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_create_app}
\alias{sms_create_app}
\title{Creates an application}
\usage{
sms_create_app(name, description, roleName, clientToken, serverGroups,
  tags)
}
\arguments{
\item{name}{Name of the new application.}

\item{description}{Description of the new application}

\item{roleName}{Name of service role in customer's account to be used by AWS SMS.}

\item{clientToken}{A unique, case-sensitive identifier you provide to ensure idempotency of
application creation.}

\item{serverGroups}{List of server groups to include in the application.}

\item{tags}{List of tags to be associated with the application.}
}
\description{
Creates an application. An application consists of one or more server
groups. Each server group contain one or more servers.
}
\section{Request syntax}{
\preformatted{sms$create_app(
  name = "string",
  description = "string",
  roleName = "string",
  clientToken = "string",
  serverGroups = list(
    list(
      serverGroupId = "string",
      name = "string",
      serverList = list(
        list(
          serverId = "string",
          serverType = "VIRTUAL_MACHINE",
          vmServer = list(
            vmServerAddress = list(
              vmManagerId = "string",
              vmId = "string"
            ),
            vmName = "string",
            vmManagerName = "string",
            vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
            vmPath = "string"
          ),
          replicationJobId = "string",
          replicationJobTerminated = TRUE|FALSE
        )
      )
    )
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
