% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app}
\alias{sms_get_app}
\title{Retrieve information about the specified application}
\usage{
sms_get_app(appId)
}
\arguments{
\item{appId}{The ID of the application.}
}
\description{
Retrieve information about the specified application.
}
\section{Request syntax}{
\preformatted{svc$get_app(
  appId = "string"
)
}
}

\keyword{internal}
