% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app_replication_configuration}
\alias{sms_get_app_replication_configuration}
\title{Retrieves the application replication configuration associated with the
specified application}
\usage{
sms_get_app_replication_configuration(appId)
}
\arguments{
\item{appId}{The ID of the application.}
}
\description{
Retrieves the application replication configuration associated with the
specified application.
}
\section{Request syntax}{
\preformatted{svc$get_app_replication_configuration(
  appId = "string"
)
}
}

\keyword{internal}
