% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app_validation_output}
\alias{sms_get_app_validation_output}
\title{Retrieves output from validating an application}
\usage{
sms_get_app_validation_output(appId)
}
\arguments{
\item{appId}{[required] The ID of the application.}
}
\description{
Retrieves output from validating an application.
}
\section{Request syntax}{
\preformatted{svc$get_app_validation_output(
  appId = "string"
)
}
}

\keyword{internal}
