% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_replication_runs}
\alias{sms_get_replication_runs}
\title{Describes the replication runs for the specified replication job}
\usage{
sms_get_replication_runs(replicationJobId, nextToken, maxResults)
}
\arguments{
\item{replicationJobId}{[required] The identifier of the replication job.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in a single call. The default
value is 50. To retrieve the remaining results, make another call with
the returned \code{NextToken} value.}
}
\description{
Describes the replication runs for the specified replication job.
}
\section{Request syntax}{
\preformatted{sms$get_replication_runs(
  replicationJobId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
