% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_launch_app}
\alias{sms_launch_app}
\title{Launches an application stack}
\usage{
sms_launch_app(appId)
}
\arguments{
\item{appId}{ID of the application to launch.}
}
\description{
Launches an application stack.
}
\section{Request syntax}{
\preformatted{sms$launch_app(
  appId = "string"
)
}
}

\keyword{internal}
