% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_start_app_replication}
\alias{sms_start_app_replication}
\title{Starts replicating an application}
\usage{
sms_start_app_replication(appId)
}
\arguments{
\item{appId}{ID of the application to replicate.}
}
\description{
Starts replicating an application.
}
\section{Request syntax}{
\preformatted{svc$start_app_replication(
  appId = "string"
)
}
}

\keyword{internal}
