% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_stop_app_replication}
\alias{sms_stop_app_replication}
\title{Stops replicating an application}
\usage{
sms_stop_app_replication(appId)
}
\arguments{
\item{appId}{ID of the application to stop replicating.}
}
\description{
Stops replicating an application.
}
\section{Request syntax}{
\preformatted{sms$stop_app_replication(
  appId = "string"
)
}
}

\keyword{internal}
