% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_terminate_app}
\alias{sms_terminate_app}
\title{Terminates the stack for an application}
\usage{
sms_terminate_app(appId)
}
\arguments{
\item{appId}{ID of the application to terminate.}
}
\description{
Terminates the stack for an application.
}
\section{Request syntax}{
\preformatted{sms$terminate_app(
  appId = "string"
)
}
}

\keyword{internal}
