% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_update_replication_job}
\alias{sms_update_replication_job}
\title{Updates the specified settings for the specified replication job}
\usage{
sms_update_replication_job(replicationJobId, frequency,
  nextReplicationRunStartTime, licenseType, roleName, description,
  numberOfRecentAmisToKeep, encrypted, kmsKeyId)
}
\arguments{
\item{replicationJobId}{[required] The identifier of the replication job.}

\item{frequency}{The time between consecutive replication runs, in hours.}

\item{nextReplicationRunStartTime}{The start time of the next replication run.}

\item{licenseType}{The license type to be used for the AMI created by a successful
replication run.}

\item{roleName}{The name of the IAM role to be used by AWS SMS.}

\item{description}{The description of the replication job.}

\item{numberOfRecentAmisToKeep}{The maximum number of SMS-created AMIs to retain. The oldest will be
deleted once the maximum number is reached and a new AMI is created.}

\item{encrypted}{When true, the replication job produces encrypted AMIs . See also
\code{KmsKeyId} below.}

\item{kmsKeyId}{KMS key ID for replication jobs that produce encrypted AMIs. Can be any
of the following:
\itemize{
\item KMS key ID
\item KMS key alias
\item ARN referring to KMS key ID
\item ARN referring to KMS key alias
}

If encrypted is \emph{true} but a KMS key id is not specified, the
customer\'s default KMS key for EBS is used.}
}
\description{
Updates the specified settings for the specified replication job.
}
\section{Request syntax}{
\preformatted{svc$update_replication_job(
  replicationJobId = "string",
  frequency = 123,
  nextReplicationRunStartTime = as.POSIXct(
    "2015-01-01"
  ),
  licenseType = "AWS"|"BYOL",
  roleName = "string",
  description = "string",
  numberOfRecentAmisToKeep = 123,
  encrypted = TRUE|FALSE,
  kmsKeyId = "string"
)
}
}

\keyword{internal}
