% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_create_address}
\alias{snowball_create_address}
\title{Creates an address for a Snowball to be shipped to}
\usage{
snowball_create_address(Address)
}
\arguments{
\item{Address}{[required] The address that you want the Snowball shipped to.}
}
\description{
Creates an address for a Snowball to be shipped to. In most regions,
addresses are validated at the time of creation. The address you provide
must be located within the serviceable area of your region. If the
address is invalid or unsupported, then an exception is thrown.
}
\section{Request syntax}{
\preformatted{svc$create_address(
  Address = list(
    AddressId = "string",
    Name = "string",
    Company = "string",
    Street1 = "string",
    Street2 = "string",
    Street3 = "string",
    City = "string",
    StateOrProvince = "string",
    PrefectureOrDistrict = "string",
    Landmark = "string",
    Country = "string",
    PostalCode = "string",
    PhoneNumber = "string",
    IsRestricted = TRUE|FALSE
  )
)
}
}

\section{Examples}{
\preformatted{# This operation creates an address for a job. Addresses are validated at
# the time of creation. The address you provide must be located within the
# serviceable area of your region. If the address is invalid or
# unsupported, then an exception is thrown.
svc$create_address(
  Address = list(
    City = "Seattle",
    Company = "My Company's Name",
    Country = "USA",
    Name = "My Name",
    PhoneNumber = "425-555-5555",
    PostalCode = "98101",
    StateOrProvince = "WA",
    Street1 = "123 Main Street"
  )
)
}
}

\keyword{internal}
