% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_describe_address}
\alias{snowball_describe_address}
\title{Takes an AddressId and returns specific details about that address in
the form of an Address object}
\usage{
snowball_describe_address(AddressId)
}
\arguments{
\item{AddressId}{[required] The automatically generated ID for a specific address.}
}
\description{
Takes an \code{AddressId} and returns specific details about that address in
the form of an \code{Address} object.
}
\section{Request syntax}{
\preformatted{snowball$describe_address(
  AddressId = "string"
)
}
}

\examples{
# This operation describes an address for a job.
\donttest{snowball$describe_address(
  AddressId = "ADID1234ab12-3eec-4eb3-9be6-9374c10eb51b"
)}

}
\keyword{internal}
