% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_get_job_manifest}
\alias{snowball_get_job_manifest}
\title{Returns a link to an Amazon S3 presigned URL for the manifest file
associated with the specified JobId value}
\usage{
snowball_get_job_manifest(JobId)
}
\arguments{
\item{JobId}{[required] The ID for a job that you want to get the manifest file for, for example
\verb{JID123e4567-e89b-12d3-a456-426655440000}.}
}
\description{
Returns a link to an Amazon S3 presigned URL for the manifest file
associated with the specified \code{JobId} value. You can access the manifest
file for up to 60 minutes after this request has been made. To access
the manifest file after 60 minutes have passed, you'll have to make
another call to the \code{GetJobManifest} action.

The manifest is an encrypted file that you can download after your job
enters the \code{WithCustomer} status. The manifest is decrypted by using the
\code{UnlockCode} code value, when you pass both values to the Snow device
through the Snowball client when the client is started for the first
time.

As a best practice, we recommend that you don't save a copy of an
\code{UnlockCode} value in the same location as the manifest file for that
job. Saving these separately helps prevent unauthorized parties from
gaining access to the Snow device associated with that job.

The credentials of a given job, including its manifest file and unlock
code, expire 90 days after the job is created.
}
\section{Request syntax}{
\preformatted{svc$get_job_manifest(
  JobId = "string"
)
}
}

\examples{
\dontrun{
# Returns a link to an Amazon S3 presigned URL for the manifest file
# associated with the specified JobId value. You can access the manifest
# file for up to 60 minutes after this request has been made. To access
# the manifest file after 60 minutes have passed, you'll have to make
# another call to the GetJobManifest action.
# 
# The manifest is an encrypted
# file that you can download after your job enters the WithCustomer
# status. The manifest is decrypted by using the UnlockCode code value,
# when you pass both values to the Snowball through the Snowball client
# when the client is started for the first time.
# 
# As a best practice, we
# recommend that you don't save a copy of an UnlockCode value in the same
# location as the manifest file for that job. Saving these separately
# helps prevent unauthorized parties from gaining access to the Snowball
# associated with that job.
# 
# The credentials of a given job, including its
# manifest file and unlock code, expire 90 days after the job is created.
svc$get_job_manifest(
  JobId = "JID123e4567-e89b-12d3-a456-426655440000"
)
}

}
\keyword{internal}
