% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_list_compatible_images}
\alias{snowball_list_compatible_images}
\title{This action returns a list of the different Amazon EC2 Amazon Machine
Images (AMIs) that are owned by your AWS account that would be supported
for use on EDGE, EDGE_C, and EDGE_CG devices}
\usage{
snowball_list_compatible_images(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results for the list of compatible images.
Currently, each supported device can store 10 AMIs.}

\item{NextToken}{HTTP requests are stateless. To identify what object comes "next" in
the list of compatible images, you can specify a value for \code{NextToken}
as the starting point for your list of returned images.}
}
\description{
This action returns a list of the different Amazon EC2 Amazon Machine
Images (AMIs) that are owned by your AWS account that would be supported
for use on \code{EDGE}, \code{EDGE_C}, and \code{EDGE_CG} devices. For more information
on compatible AMIs, see \href{http://docs.aws.amazon.com/snowball/latest/developer-guide/using-ec2.html}{Using Amazon EC2 Compute Instances}
in the \emph{AWS Snowball Developer Guide}.
}
\section{Request syntax}{
\preformatted{snowball$list_compatible_images(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
