% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_update_job}
\alias{snowball_update_job}
\title{While a job's JobState value is New, you can update some of the
information associated with a job}
\usage{
snowball_update_job(JobId, RoleARN, Notification, Resources, AddressId,
  ShippingOption, Description, SnowballCapacityPreference,
  ForwardingAddressId)
}
\arguments{
\item{JobId}{[required] The job ID of the job that you want to update, for example
\verb{JID123e4567-e89b-12d3-a456-426655440000}.}

\item{RoleARN}{The new role Amazon Resource Name (ARN) that you want to associate with
this job. To create a role ARN, use the
\href{https://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateRole.html}{CreateRole}AWS
Identity and Access Management (IAM) API action.}

\item{Notification}{The new or updated Notification object.}

\item{Resources}{The updated \code{JobResource} object, or the updated JobResource object.}

\item{AddressId}{The ID of the updated Address object.}

\item{ShippingOption}{The updated shipping option value of this job\'s ShippingDetails object.}

\item{Description}{The updated description of this job\'s JobMetadata object.}

\item{SnowballCapacityPreference}{The updated \code{SnowballCapacityPreference} of this job\'s JobMetadata
object. The 50 TB Snowballs are only available in the US regions.}

\item{ForwardingAddressId}{The updated ID for the forwarding address for a job. This field is not
supported in most regions.}
}
\description{
While a job\'s \code{JobState} value is \code{New}, you can update some of the
information associated with a job. Once the job changes to a different
job state, usually within 60 minutes of the job being created, this
action is no longer available.
}
\section{Request syntax}{
\preformatted{svc$update_job(
  JobId = "string",
  RoleARN = "string",
  Notification = list(
    SnsTopicARN = "string",
    JobStatesToNotify = list(
      "New"|"PreparingAppliance"|"PreparingShipment"|"InTransitToCustomer"|"WithCustomer"|"InTransitToAWS"|"WithAWSSortingFacility"|"WithAWS"|"InProgress"|"Complete"|"Cancelled"|"Listing"|"Pending"
    ),
    NotifyAll = TRUE|FALSE
  ),
  Resources = list(
    S3Resources = list(
      list(
        BucketArn = "string",
        KeyRange = list(
          BeginMarker = "string",
          EndMarker = "string"
        )
      )
    ),
    LambdaResources = list(
      list(
        LambdaArn = "string",
        EventTriggers = list(
          list(
            EventResourceARN = "string"
          )
        )
      )
    ),
    Ec2AmiResources = list(
      list(
        AmiId = "string",
        SnowballAmiId = "string"
      )
    )
  ),
  AddressId = "string",
  ShippingOption = "SECOND_DAY"|"NEXT_DAY"|"EXPRESS"|"STANDARD",
  Description = "string",
  SnowballCapacityPreference = "T50"|"T80"|"T100"|"T42"|"T98"|"T8"|"NoPreference",
  ForwardingAddressId = "string"
)
}
}

\examples{
\dontrun{
# This action allows you to update certain parameters for a job. Once the
# job changes to a different job state, usually within 60 minutes of the
# job being created, this action is no longer available.
svc$update_job(
  AddressId = "ADID1234ab12-3eec-4eb3-9be6-9374c10eb51b",
  Description = "Upgraded to Edge, shipped to Finance Dept, and requested faster shipping sp...",
  JobId = "JID123e4567-e89b-12d3-a456-426655440000",
  ShippingOption = "NEXT_DAY",
  SnowballCapacityPreference = "T100"
)
}

}
\keyword{internal}
