% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_update_job_shipment_state}
\alias{snowball_update_job_shipment_state}
\title{Updates the state when a the shipment states changes to a different
state}
\usage{
snowball_update_job_shipment_state(JobId, ShipmentState)
}
\arguments{
\item{JobId}{[required] The job ID of the job whose shipment date you want to update, for
example \verb{JID123e4567-e89b-12d3-a456-426655440000}.}

\item{ShipmentState}{[required] The state of a device when it is being shipped.

Set to \code{RECEIVED} when the device arrives at your location.

Set to \code{RETURNED} when you have returned the device to AWS.}
}
\description{
Updates the state when a the shipment states changes to a different
state.
}
\section{Request syntax}{
\preformatted{svc$update_job_shipment_state(
  JobId = "string",
  ShipmentState = "RECEIVED"|"RETURNED"
)
}
}

\keyword{internal}
