% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_delete_platform_application}
\alias{sns_delete_platform_application}
\title{Deletes a platform application object for one of the supported push
notification services, such as APNS and FCM}
\usage{
sns_delete_platform_application(PlatformApplicationArn)
}
\arguments{
\item{PlatformApplicationArn}{[required] PlatformApplicationArn of platform application object to delete.}
}
\description{
Deletes a platform application object for one of the supported push
notification services, such as APNS and FCM. For more information, see
\href{https://docs.aws.amazon.com/sns/latest/dg/SNSMobilePush.html}{Using Amazon SNS Mobile Push Notifications}.
}
\section{Request syntax}{
\preformatted{svc$delete_platform_application(
  PlatformApplicationArn = "string"
)
}
}

\keyword{internal}
