% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_get_endpoint_attributes}
\alias{sns_get_endpoint_attributes}
\title{Retrieves the endpoint attributes for a device on one of the supported
push notification services, such as FCM and APNS}
\usage{
sns_get_endpoint_attributes(EndpointArn)
}
\arguments{
\item{EndpointArn}{[required] EndpointArn for GetEndpointAttributes input.}
}
\description{
Retrieves the endpoint attributes for a device on one of the supported
push notification services, such as FCM and APNS. For more information,
see \href{https://docs.aws.amazon.com/sns/latest/dg/SNSMobilePush.html}{Using Amazon SNS Mobile Push Notifications}.
}
\section{Request syntax}{
\preformatted{svc$get_endpoint_attributes(
  EndpointArn = "string"
)
}
}

\keyword{internal}
