% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_get_topic_attributes}
\alias{sns_get_topic_attributes}
\title{Returns all of the properties of a topic}
\usage{
sns_get_topic_attributes(TopicArn)
}
\arguments{
\item{TopicArn}{[required] The ARN of the topic whose properties you want to get.}
}
\description{
Returns all of the properties of a topic. Topic properties returned
might differ based on the authorization of the user.
}
\section{Request syntax}{
\preformatted{svc$get_topic_attributes(
  TopicArn = "string"
)
}
}

\keyword{internal}
