% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_list_topics}
\alias{sns_list_topics}
\title{Returns a list of the requester's topics}
\usage{
sns_list_topics(NextToken)
}
\arguments{
\item{NextToken}{Token returned by the previous \code{ListTopics} request.}
}
\description{
Returns a list of the requester's topics. Each call returns a limited
list of topics, up to 100. If there are more topics, a \code{NextToken} is
also returned. Use the \code{NextToken} parameter in a new \code{ListTopics} call
to get further results.

This action is throttled at 30 transactions per second (TPS).
}
\section{Request syntax}{
\preformatted{svc$list_topics(
  NextToken = "string"
)
}
}

\keyword{internal}
