% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_create_ops_item}
\alias{ssm_create_ops_item}
\title{Creates a new OpsItem}
\usage{
ssm_create_ops_item(Description, OperationalData, Notifications,
  Priority, RelatedOpsItems, Source, Title, Tags)
}
\arguments{
\item{Description}{[required] Information about the OpsItem.}

\item{OperationalData}{Operational data is custom data that provides useful reference details
about the OpsItem. For example, you can specify log files, error
strings, license keys, troubleshooting tips, or other relevant data. You
enter operational data as key-value pairs. The key has a maximum length
of 128 characters. The value has a maximum size of 20 KB.

This custom data is searchable, but with restrictions. For the
\code{Searchable operational data} feature, all users with access to the
OpsItem Overview page (as provided by the DescribeOpsItems API action)
can view and search on the specified data. For the
\code{Private operational data} feature, the data is only viewable by users
who have access to the OpsItem (as provided by the GetOpsItem API
action).}

\item{Notifications}{The Amazon Resource Name (ARN) of an SNS topic where notifications are
sent when this OpsItem is edited or changed.}

\item{Priority}{The importance of this OpsItem in relation to other OpsItems in the
system.}

\item{RelatedOpsItems}{One or more OpsItems that share something in common with the current
OpsItems. For example, related OpsItems can include OpsItems with
similar error messages, impacted resources, or statuses for the impacted
resource.}

\item{Source}{[required] The origin of the OpsItem, such as Amazon EC2 or AWS Systems Manager.}

\item{Title}{[required] A short heading that describes the nature of the OpsItem and the
impacted resource.}

\item{Tags}{Optional metadata that you assign to a resource. Tags enable you to
categorize a resource in different ways, such as by purpose, owner, or
environment. For example, you might want to tag an OpsItem to identify
the AWS resource or the type of issue. In this case, you could specify
the following key name/value pairs:
\itemize{
\item \code{Key=source,Value=EC2-instance}
\item \code{Key=status,Value=stopped}
}

To add tags to an existing OpsItem, use the AddTagsToResource action.}
}
\description{
Creates a new OpsItem. You must have permission in AWS Identity and
Access Management (IAM) to create a new OpsItem. For more information,
see \href{http://docs.aws.amazon.com/systems-manager/latest/userguide/OpsItems-getting-started.html}{Getting Started with OpsItems}
in the \emph{AWS Systems Manager User Guide}.
}
\details{
Operations engineers and IT professionals use the Systems Manager
OpsItems capability to view, investigate, and remediate operational
issues impacting the performance and health of their AWS resources. For
more information, see \href{http://docs.aws.amazon.com/systems-manager/latest/userguide/OpsItems.html}{AWS Systems Manager OpsItems}
in the \emph{AWS Systems Manager User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_ops_item(
  Description = "string",
  OperationalData = list(
    list(
      Value = "string",
      Type = "SearchableString"|"String"
    )
  ),
  Notifications = list(
    list(
      Arn = "string"
    )
  ),
  Priority = 123,
  RelatedOpsItems = list(
    list(
      OpsItemId = "string"
    )
  ),
  Source = "string",
  Title = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
