% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_create_ops_metadata}
\alias{ssm_create_ops_metadata}
\title{If you create a new application in Application Manager, Systems Manager
calls this API action to specify information about the new application,
including the application type}
\usage{
ssm_create_ops_metadata(ResourceId, Metadata)
}
\arguments{
\item{ResourceId}{[required] A resource ID for a new Application Manager application.}

\item{Metadata}{Metadata for a new Application Manager application.}
}
\description{
If you create a new application in Application Manager, Systems Manager
calls this API action to specify information about the new application,
including the application type.
}
\section{Request syntax}{
\preformatted{svc$create_ops_metadata(
  ResourceId = "string",
  Metadata = list(
    list(
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
