% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_maintenance_window}
\alias{ssm_delete_maintenance_window}
\title{Deletes a Maintenance Window}
\usage{
ssm_delete_maintenance_window(WindowId)
}
\arguments{
\item{WindowId}{[required] The ID of the Maintenance Window to delete.}
}
\description{
Deletes a Maintenance Window.
}
\section{Request syntax}{
\preformatted{ssm$delete_maintenance_window(
  WindowId = "string"
)
}
}

\keyword{internal}
