% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_parameter}
\alias{ssm_delete_parameter}
\title{Delete a parameter from the system}
\usage{
ssm_delete_parameter(Name)
}
\arguments{
\item{Name}{[required] The name of the parameter to delete.}
}
\description{
Delete a parameter from the system.
}
\section{Request syntax}{
\preformatted{ssm$delete_parameter(
  Name = "string"
)
}
}

\keyword{internal}
