% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_deregister_task_from_maintenance_window}
\alias{ssm_deregister_task_from_maintenance_window}
\title{Removes a task from a Maintenance Window}
\usage{
ssm_deregister_task_from_maintenance_window(WindowId, WindowTaskId)
}
\arguments{
\item{WindowId}{[required] The ID of the Maintenance Window the task should be removed from.}

\item{WindowTaskId}{[required] The ID of the task to remove from the Maintenance Window.}
}
\description{
Removes a task from a Maintenance Window.
}
\section{Request syntax}{
\preformatted{svc$deregister_task_from_maintenance_window(
  WindowId = "string",
  WindowTaskId = "string"
)
}
}

\keyword{internal}
