% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_instance_patch_states}
\alias{ssm_describe_instance_patch_states}
\title{Retrieves the high-level patch state of one or more instances}
\usage{
ssm_describe_instance_patch_states(InstanceIds, NextToken, MaxResults)
}
\arguments{
\item{InstanceIds}{[required] The ID of the instance whose patch state information should be
retrieved.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxResults}{The maximum number of instances to return (per page).}
}
\description{
Retrieves the high-level patch state of one or more instances.
}
\section{Request syntax}{
\preformatted{svc$describe_instance_patch_states(
  InstanceIds = list(
    "string"
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
