% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_automation_execution}
\alias{ssm_get_automation_execution}
\title{Get detailed information about a particular Automation execution}
\usage{
ssm_get_automation_execution(AutomationExecutionId)
}
\arguments{
\item{AutomationExecutionId}{[required] The unique identifier for an existing automation execution to examine.
The execution ID is returned by StartAutomationExecution when the
execution of an Automation document is initiated.}
}
\description{
Get detailed information about a particular Automation execution.
}
\section{Request syntax}{
\preformatted{ssm$get_automation_execution(
  AutomationExecutionId = "string"
)
}
}

\keyword{internal}
