% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_default_patch_baseline}
\alias{ssm_get_default_patch_baseline}
\title{Retrieves the default patch baseline}
\usage{
ssm_get_default_patch_baseline(OperatingSystem)
}
\arguments{
\item{OperatingSystem}{Returns the default patch baseline for the specified operating system.}
}
\description{
Retrieves the default patch baseline. Note that Systems Manager supports
creating multiple default patch baselines. For example, you can create a
default patch baseline for each operating system.
}
\details{
If you do not specify an operating system value, the default patch
baseline for Windows is returned.
}
\section{Request syntax}{
\preformatted{svc$get_default_patch_baseline(
  OperatingSystem = "WINDOWS"|"AMAZON_LINUX"|"AMAZON_LINUX_2"|"UBUNTU"|"REDHAT_ENTERPRISE_LINUX"|"SUSE"|"CENTOS"|"ORACLE_LINUX"|"DEBIAN"
)
}
}

\keyword{internal}
