% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_maintenance_window_execution_task_invocation}
\alias{ssm_get_maintenance_window_execution_task_invocation}
\title{Retrieves a task invocation}
\usage{
ssm_get_maintenance_window_execution_task_invocation(WindowExecutionId,
  TaskId, InvocationId)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the maintenance window execution for which the task is a part.}

\item{TaskId}{[required] The ID of the specific task in the maintenance window task that should
be retrieved.}

\item{InvocationId}{[required] The invocation ID to retrieve.}
}
\description{
Retrieves a task invocation. A task invocation is a specific task
running on a specific target. maintenance windows report status for all
invocations.
}
\section{Request syntax}{
\preformatted{svc$get_maintenance_window_execution_task_invocation(
  WindowExecutionId = "string",
  TaskId = "string",
  InvocationId = "string"
)
}
}

\keyword{internal}
