% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_parameter}
\alias{ssm_get_parameter}
\title{Get information about a parameter by using the parameter name}
\usage{
ssm_get_parameter(Name, WithDecryption)
}
\arguments{
\item{Name}{[required] The name of the parameter you want to query.}

\item{WithDecryption}{Return decrypted values for secure string parameters. This flag is
ignored for String and StringList parameter types.}
}
\description{
Get information about a parameter by using the parameter name. Don\'t
confuse this API action with the GetParameters API action.
}
\section{Request syntax}{
\preformatted{svc$get_parameter(
  Name = "string",
  WithDecryption = TRUE|FALSE
)
}
}

\keyword{internal}
