% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_resume_session}
\alias{ssm_resume_session}
\title{Reconnects a session to an instance after it has been disconnected}
\usage{
ssm_resume_session(SessionId)
}
\arguments{
\item{SessionId}{[required] The ID of the disconnected session to resume.}
}
\description{
Reconnects a session to an instance after it has been disconnected.
Connections can be resumed for disconnected sessions, but not terminated
sessions.

This command is primarily for use by client machines to automatically
reconnect during intermittent network issues. It is not intended for any
other use.
}
\section{Request syntax}{
\preformatted{svc$resume_session(
  SessionId = "string"
)
}
}

\keyword{internal}
