% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_update_association_status}
\alias{ssm_update_association_status}
\title{Updates the status of the Systems Manager document associated with the
specified instance}
\usage{
ssm_update_association_status(Name, InstanceId, AssociationStatus)
}
\arguments{
\item{Name}{[required] The name of the Systems Manager document.}

\item{InstanceId}{[required] The ID of the instance.}

\item{AssociationStatus}{[required] The association status.}
}
\description{
Updates the status of the Systems Manager document associated with the
specified instance.
}
\section{Request syntax}{
\preformatted{ssm$update_association_status(
  Name = "string",
  InstanceId = "string",
  AssociationStatus = list(
    Date = as.POSIXct(
      "2015-01-01"
    ),
    Name = "Pending"|"Success"|"Failed",
    Message = "string",
    AdditionalInfo = "string"
  )
)
}
}

\keyword{internal}
