% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_create_account_assignment}
\alias{ssoadmin_create_account_assignment}
\title{Assigns access to a principal for a specified AWS account using a
specified permission set}
\usage{
ssoadmin_create_account_assignment(InstanceArn, TargetId, TargetType,
  PermissionSetArn, PrincipalType, PrincipalId)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{TargetId}{[required] TargetID is an AWS account identifier, typically a 10-12 digit string
(For example, 123456789012).}

\item{TargetType}{[required] The entity type for which the assignment will be created.}

\item{PermissionSetArn}{[required] The ARN of the permission set that the admin wants to grant the
principal access to.}

\item{PrincipalType}{[required] The entity type for which the assignment will be created.}

\item{PrincipalId}{[required] An identifier for an object in AWS SSO, such as a user or group.
PrincipalIds are GUIDs (For example,
f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about
PrincipalIds in AWS SSO, see the AWS SSO Identity Store API Reference.}
}
\description{
Assigns access to a principal for a specified AWS account using a
specified permission set.
}
\details{
The term \emph{principal} here refers to a user or group that is defined in
AWS SSO.

As part of a successful \code{CreateAccountAssignment} call, the specified
permission set will automatically be provisioned to the account in the
form of an IAM policy attached to the SSO-created IAM role. If the
permission set is subsequently updated, the corresponding IAM policies
attached to roles in your accounts will not be updated automatically. In
this case, you will need to call \verb{<a>ProvisionPermissionSet</a>} to
make these updates.
}
\section{Request syntax}{
\preformatted{svc$create_account_assignment(
  InstanceArn = "string",
  TargetId = "string",
  TargetType = "AWS_ACCOUNT",
  PermissionSetArn = "string",
  PrincipalType = "USER"|"GROUP",
  PrincipalId = "string"
)
}
}

\keyword{internal}
