% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_create_permission_set}
\alias{ssoadmin_create_permission_set}
\title{Creates a permission set within a specified SSO instance}
\usage{
ssoadmin_create_permission_set(Name, Description, InstanceArn,
  SessionDuration, RelayState, Tags)
}
\arguments{
\item{Name}{[required] The name of the PermissionSet.}

\item{Description}{The description of the PermissionSet.}

\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{SessionDuration}{The length of time that the application user sessions are valid in the
ISO-8601 standard.}

\item{RelayState}{Used to redirect users within the application during the federation
authentication process.}

\item{Tags}{The tags to attach to the new PermissionSet.}
}
\description{
Creates a permission set within a specified SSO instance.

To grant users and groups access to AWS account resources, use
\verb{<a>CreateAccountAssignment</a>}.
}
\section{Request syntax}{
\preformatted{svc$create_permission_set(
  Name = "string",
  Description = "string",
  InstanceArn = "string",
  SessionDuration = "string",
  RelayState = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
