% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_account_assignment_deletion_status}
\alias{ssoadmin_describe_account_assignment_deletion_status}
\title{Describes the status of the assignment deletion request}
\usage{
ssoadmin_describe_account_assignment_deletion_status(InstanceArn,
  AccountAssignmentDeletionRequestId)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{AccountAssignmentDeletionRequestId}{[required] The identifier that is used to track the request operation progress.}
}
\description{
Describes the status of the assignment deletion request.
}
\section{Request syntax}{
\preformatted{svc$describe_account_assignment_deletion_status(
  InstanceArn = "string",
  AccountAssignmentDeletionRequestId = "string"
)
}
}

\keyword{internal}
