% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_instance_access_control_attribute_configuration}
\alias{ssoadmin_describe_instance_access_control_attribute_configuration}
\title{Returns the list of AWS SSO identity store attributes that have been
configured to work with attributes-based access control (ABAC) for the
specified AWS SSO instance}
\usage{
ssoadmin_describe_instance_access_control_attribute_configuration(
  InstanceArn)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.}
}
\description{
Returns the list of AWS SSO identity store attributes that have been
configured to work with attributes-based access control (ABAC) for the
specified AWS SSO instance. This will not return attributes configured
and sent by an external identity provider. For more information about
ABAC, see Attribute-Based Access Control in the \emph{AWS SSO User Guide}.
}
\section{Request syntax}{
\preformatted{svc$describe_instance_access_control_attribute_configuration(
  InstanceArn = "string"
)
}
}

\keyword{internal}
