% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_get_inline_policy_for_permission_set}
\alias{ssoadmin_get_inline_policy_for_permission_set}
\title{Obtains the inline policy assigned to the permission set}
\usage{
ssoadmin_get_inline_policy_for_permission_set(InstanceArn,
  PermissionSetArn)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{PermissionSetArn}{[required] The ARN of the permission set.}
}
\description{
Obtains the inline policy assigned to the permission set.
}
\section{Request syntax}{
\preformatted{svc$get_inline_policy_for_permission_set(
  InstanceArn = "string",
  PermissionSetArn = "string"
)
}
}

\keyword{internal}
