% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_permission_set_provisioning_status}
\alias{ssoadmin_list_permission_set_provisioning_status}
\title{Lists the status of the permission set provisioning requests for a
specified SSO instance}
\usage{
ssoadmin_list_permission_set_provisioning_status(InstanceArn,
  MaxResults, NextToken, Filter)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{MaxResults}{The maximum number of results to display for the assignment.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}

\item{Filter}{Filters results based on the passed attribute value.}
}
\description{
Lists the status of the permission set provisioning requests for a
specified SSO instance.
}
\section{Request syntax}{
\preformatted{svc$list_permission_set_provisioning_status(
  InstanceArn = "string",
  MaxResults = 123,
  NextToken = "string",
  Filter = list(
    Status = "IN_PROGRESS"|"FAILED"|"SUCCEEDED"
  )
)
}
}

\keyword{internal}
