% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_permission_sets}
\alias{ssoadmin_list_permission_sets}
\title{Lists the PermissionSets in an SSO instance}
\usage{
ssoadmin_list_permission_sets(InstanceArn, NextToken, MaxResults)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}

\item{MaxResults}{The maximum number of results to display for the assignment.}
}
\description{
Lists the PermissionSets in an SSO instance.
}
\section{Request syntax}{
\preformatted{svc$list_permission_sets(
  InstanceArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
