% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_provision_permission_set}
\alias{ssoadmin_provision_permission_set}
\title{The process by which a specified permission set is provisioned to the
specified target}
\usage{
ssoadmin_provision_permission_set(InstanceArn, PermissionSetArn,
  TargetId, TargetType)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{PermissionSetArn}{[required] The ARN of the permission set.}

\item{TargetId}{TargetID is an AWS account identifier, typically a 10-12 digit string
(For example, 123456789012).}

\item{TargetType}{[required] The entity type for which the assignment will be created.}
}
\description{
The process by which a specified permission set is provisioned to the
specified target.
}
\section{Request syntax}{
\preformatted{svc$provision_permission_set(
  InstanceArn = "string",
  PermissionSetArn = "string",
  TargetId = "string",
  TargetType = "AWS_ACCOUNT"|"ALL_PROVISIONED_ACCOUNTS"
)
}
}

\keyword{internal}
