% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_untag_resource}
\alias{ssoadmin_untag_resource}
\title{Disassociates a set of tags from a specified resource}
\usage{
ssoadmin_untag_resource(InstanceArn, ResourceArn, TagKeys)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{ResourceArn}{[required] The ARN of the resource with the tags to be listed.}

\item{TagKeys}{[required] The keys of tags that are attached to the resource.}
}
\description{
Disassociates a set of tags from a specified resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  InstanceArn = "string",
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
