% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_update_permission_set}
\alias{ssoadmin_update_permission_set}
\title{Updates an existing permission set}
\usage{
ssoadmin_update_permission_set(InstanceArn, PermissionSetArn,
  Description, SessionDuration, RelayState)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{PermissionSetArn}{[required] The ARN of the permission set.}

\item{Description}{The description of the PermissionSet.}

\item{SessionDuration}{The length of time that the application user sessions are valid for in
the ISO-8601 standard.}

\item{RelayState}{Used to redirect users within the application during the federation
authentication process.}
}
\description{
Updates an existing permission set.
}
\section{Request syntax}{
\preformatted{svc$update_permission_set(
  InstanceArn = "string",
  PermissionSetArn = "string",
  Description = "string",
  SessionDuration = "string",
  RelayState = "string"
)
}
}

\keyword{internal}
