% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_create_smb_file_share}
\alias{storagegateway_create_smb_file_share}
\title{Creates a Server Message Block (SMB) file share on an existing file
gateway}
\usage{
storagegateway_create_smb_file_share(ClientToken, GatewayARN,
  KMSEncrypted, KMSKey, Role, LocationARN, DefaultStorageClass, ObjectACL,
  ReadOnly, GuessMIMETypeEnabled, RequesterPays, ValidUserList,
  InvalidUserList, Authentication)
}
\arguments{
\item{ClientToken}{[required] A unique string value that you supply that is used by file gateway to
ensure idempotent file share creation.}

\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the file gateway on which you want to
create a file share.}

\item{KMSEncrypted}{True to use Amazon S3 server side encryption with your own AWS KMS key,
or false to use a key managed by Amazon S3. Optional.}

\item{KMSKey}{The Amazon Resource Name (ARN) of the AWS KMS key used for Amazon S3
server side encryption. This value can only be set when KMSEncrypted is
true. Optional.}

\item{Role}{[required] The ARN of the AWS Identity and Access Management (IAM) role that a file
gateway assumes when it accesses the underlying storage.}

\item{LocationARN}{[required] The ARN of the backed storage used for storing file data.}

\item{DefaultStorageClass}{The default storage class for objects put into an Amazon S3 bucket by
the file gateway. Possible values are \code{S3_STANDARD}, \code{S3_STANDARD_IA},
or \code{S3_ONEZONE_IA}. If this field is not populated, the default value
\code{S3_STANDARD} is used. Optional.}

\item{ObjectACL}{A value that sets the access control list permission for objects in the
S3 bucket that a file gateway puts objects into. The default value is
"private".}

\item{ReadOnly}{A value that sets the write status of a file share. This value is true
if the write status is read-only, and otherwise false.}

\item{GuessMIMETypeEnabled}{A value that enables guessing of the MIME type for uploaded objects
based on file extensions. Set this value to true to enable MIME type
guessing, and otherwise to false. The default value is true.}

\item{RequesterPays}{A value that sets the access control list permission for objects in the
Amazon S3 bucket that a file gateway puts objects into. The default
value is \code{private}.}

\item{ValidUserList}{A list of users or groups in the Active Directory that are allowed to
access the file share. A group must be prefixed with the @ character.
For example \code{@group1}. Can only be set if Authentication is set to
\code{ActiveDirectory}.}

\item{InvalidUserList}{A list of users or groups in the Active Directory that are not allowed
to access the file share. A group must be prefixed with the @ character.
For example \code{@group1}. Can only be set if Authentication is set to
\code{ActiveDirectory}.}

\item{Authentication}{The authentication method that users use to access the file share.

Valid values are \code{ActiveDirectory} or \code{GuestAccess}. The default is
\code{ActiveDirectory}.}
}
\description{
Creates a Server Message Block (SMB) file share on an existing file
gateway. In Storage Gateway, a file share is a file system mount point
backed by Amazon S3 cloud storage. Storage Gateway expose file shares
using a SMB interface. This operation is only supported for file
gateways.
}
\details{
File gateways require AWS Security Token Service (AWS STS) to be
activated to enable you to create a file share. Make sure that AWS STS
is activated in the AWS Region you are creating your file gateway in. If
AWS STS is not activated in this AWS Region, activate it. For
information about how to activate AWS STS, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_temp_enable-regions.html}{Activating and Deactivating AWS STS in an AWS Region}
in the \emph{AWS Identity and Access Management User Guide.}

File gateways don't support creating hard or symbolic links on a file
share.
}
\section{Request syntax}{
\preformatted{storagegateway$create_smb_file_share(
  ClientToken = "string",
  GatewayARN = "string",
  KMSEncrypted = TRUE|FALSE,
  KMSKey = "string",
  Role = "string",
  LocationARN = "string",
  DefaultStorageClass = "string",
  ObjectACL = "private"|"public-read"|"public-read-write"|"authenticated-read"|"bucket-owner-read"|"bucket-owner-full-control"|"aws-exec-read",
  ReadOnly = TRUE|FALSE,
  GuessMIMETypeEnabled = TRUE|FALSE,
  RequesterPays = TRUE|FALSE,
  ValidUserList = list(
    "string"
  ),
  InvalidUserList = list(
    "string"
  ),
  Authentication = "string"
)
}
}

\keyword{internal}
