% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_create_tape_with_barcode}
\alias{storagegateway_create_tape_with_barcode}
\title{Creates a virtual tape by using your own barcode}
\usage{
storagegateway_create_tape_with_barcode(GatewayARN, TapeSizeInBytes,
  TapeBarcode, KMSEncrypted, KMSKey)
}
\arguments{
\item{GatewayARN}{[required] The unique Amazon Resource Name (ARN) that represents the gateway to
associate the virtual tape with. Use the ListGateways operation to
return a list of gateways for your account and region.}

\item{TapeSizeInBytes}{[required] The size, in bytes, of the virtual tape that you want to create.

The size must be aligned by gigabyte (1024*1024*1024 byte).}

\item{TapeBarcode}{[required] The barcode that you want to assign to the tape.

Barcodes cannot be reused. This includes barcodes used for tapes that
have been deleted.}

\item{KMSEncrypted}{True to use Amazon S3 server side encryption with your own AWS KMS key,
or false to use a key managed by Amazon S3. Optional.}

\item{KMSKey}{The Amazon Resource Name (ARN) of the AWS KMS Key used for Amazon S3
server side encryption. This value can only be set when KMSEncrypted is
true. Optional.}
}
\description{
Creates a virtual tape by using your own barcode. You write data to the
virtual tape and then archive the tape. A barcode is unique and can not
be reused if it has already been used on a tape . This applies to
barcodes used on deleted tapes. This operation is only supported in the
tape gateway type.
}
\details{
Cache storage must be allocated to the gateway before you can create a
virtual tape. Use the AddCache operation to add cache storage to a
gateway.
}
\section{Request syntax}{
\preformatted{storagegateway$create_tape_with_barcode(
  GatewayARN = "string",
  TapeSizeInBytes = 123,
  TapeBarcode = "string",
  KMSEncrypted = TRUE|FALSE,
  KMSKey = "string"
)
}
}

\examples{
# Creates a virtual tape by using your own barcode.
\donttest{storagegateway$create_tape_with_barcode(
  GatewayARN = "arn:aws:storagegateway:us-east-1:999999999999:gateway/sgw-12A3456B",
  TapeBarcode = "TEST12345",
  TapeSizeInBytes = 107374182400
)}

}
\keyword{internal}
