% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_update_smb_file_share}
\alias{storagegateway_update_smb_file_share}
\title{Updates a Server Message Block (SMB) file share}
\usage{
storagegateway_update_smb_file_share(FileShareARN, KMSEncrypted, KMSKey,
  DefaultStorageClass, ObjectACL, ReadOnly, GuessMIMETypeEnabled,
  RequesterPays, ValidUserList, InvalidUserList)
}
\arguments{
\item{FileShareARN}{[required] The Amazon Resource Name (ARN) of the SMB file share that you want to
update.}

\item{KMSEncrypted}{True to use Amazon S3 server side encryption with your own AWS KMS key,
or false to use a key managed by Amazon S3. Optional.}

\item{KMSKey}{The Amazon Resource Name (ARN) of the AWS KMS key used for Amazon S3
server side encryption. This value can only be set when KMSEncrypted is
true. Optional.}

\item{DefaultStorageClass}{The default storage class for objects put into an Amazon S3 bucket by
the file gateway. Possible values are \code{S3_STANDARD}, \code{S3_STANDARD_IA},
or \code{S3_ONEZONE_IA}. If this field is not populated, the default value
\code{S3_STANDARD} is used. Optional.}

\item{ObjectACL}{A value that sets the access control list permission for objects in the
S3 bucket that a file gateway puts objects into. The default value is
"private".}

\item{ReadOnly}{A value that sets the write status of a file share. This value is true
if the write status is read-only, and otherwise false.}

\item{GuessMIMETypeEnabled}{A value that enables guessing of the MIME type for uploaded objects
based on file extensions. Set this value to true to enable MIME type
guessing, and otherwise to false. The default value is true.}

\item{RequesterPays}{A value that sets the access control list permission for objects in the
Amazon S3 bucket that a file gateway puts objects into. The default
value is \code{private}.}

\item{ValidUserList}{A list of users or groups in the Active Directory that are allowed to
access the file share. A group must be prefixed with the @ character.
For example \code{@group1}. Can only be set if Authentication is set to
\code{ActiveDirectory}.}

\item{InvalidUserList}{A list of users or groups in the Active Directory that are not allowed
to access the file share. A group must be prefixed with the @ character.
For example \code{@group1}. Can only be set if Authentication is set to
\code{ActiveDirectory}.}
}
\description{
Updates a Server Message Block (SMB) file share.
}
\details{
To leave a file share field unchanged, set the corresponding input field
to null. This operation is only supported for file gateways.

File gateways require AWS Security Token Service (AWS STS) to be
activated to enable you to create a file share. Make sure that AWS STS
is activated in the AWS Region you are creating your file gateway in. If
AWS STS is not activated in this AWS Region, activate it. For
information about how to activate AWS STS, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_temp_enable-regions.html}{Activating and Deactivating AWS STS in an AWS Region}
in the \emph{AWS Identity and Access Management User Guide.}

File gateways don't support creating hard or symbolic links on a file
share.
}
\section{Request syntax}{
\preformatted{svc$update_smb_file_share(
  FileShareARN = "string",
  KMSEncrypted = TRUE|FALSE,
  KMSKey = "string",
  DefaultStorageClass = "string",
  ObjectACL = "private"|"public-read"|"public-read-write"|"authenticated-read"|"bucket-owner-read"|"bucket-owner-full-control"|"aws-exec-read",
  ReadOnly = TRUE|FALSE,
  GuessMIMETypeEnabled = TRUE|FALSE,
  RequesterPays = TRUE|FALSE,
  ValidUserList = list(
    "string"
  ),
  InvalidUserList = list(
    "string"
  )
)
}
}

\keyword{internal}
