% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_severity_levels}
\alias{support_describe_severity_levels}
\title{Returns the list of severity levels that you can assign to an AWS
Support case}
\usage{
support_describe_severity_levels(language)
}
\arguments{
\item{language}{The ISO 639-1 code for the language in which AWS provides support. AWS
Support currently supports English (\"en\") and Japanese (\"ja\").
Language parameters must be passed explicitly for operations that take
them.}
}
\description{
Returns the list of severity levels that you can assign to an AWS
Support case. The severity level for a case is also a field in the
CaseDetails data type that you include for a CreateCase request.
}
\details{
\itemize{
\item You must have a Business or Enterprise support plan to use the AWS
Support API.
\item If you call the AWS Support API from an account that does not have a
Business or Enterprise support plan, the
\code{SubscriptionRequiredException} error message appears. For
information about changing your support plan, see \href{http://aws.amazon.com/premiumsupport/}{AWS Support}.
}
}
\section{Request syntax}{
\preformatted{svc$describe_severity_levels(
  language = "string"
)
}
}

\keyword{internal}
