% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_trusted_advisor_checks}
\alias{support_describe_trusted_advisor_checks}
\title{Returns information about all available Trusted Advisor checks,
including name, ID, category, description, and metadata}
\usage{
support_describe_trusted_advisor_checks(language)
}
\arguments{
\item{language}{[required] The ISO 639-1 code for the language in which AWS provides support. AWS
Support currently supports English ("en") and Japanese ("ja").
Language parameters must be passed explicitly for operations that take
them.}
}
\description{
Returns information about all available Trusted Advisor checks,
including name, ID, category, description, and metadata. You must
specify a language code; English ("en") and Japanese ("ja") are
currently supported. The response contains a
TrustedAdvisorCheckDescription for each check.
}
\section{Request syntax}{
\preformatted{support$describe_trusted_advisor_checks(
  language = "string"
)
}
}

\keyword{internal}
