% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_create_canary}
\alias{synthetics_create_canary}
\title{Creates a canary}
\usage{
synthetics_create_canary(Name, Code, ArtifactS3Location,
  ExecutionRoleArn, Schedule, RunConfig, SuccessRetentionPeriodInDays,
  FailureRetentionPeriodInDays, RuntimeVersion, VpcConfig, Tags)
}
\arguments{
\item{Name}{[required] The name for this canary. Be sure to give it a descriptive name that
distinguishes it from other canaries in your account.

Do not include secrets or proprietary information in your canary names.
The canary name makes up part of the canary ARN, and the ARN is included
in outbound calls over the internet. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/servicelens_canaries_security.html}{Security Considerations for Synthetics Canaries}.}

\item{Code}{[required] A structure that includes the entry point from which the canary should
start running your script. If the script is stored in an S3 bucket, the
bucket name, key, and version are also included.}

\item{ArtifactS3Location}{[required] The location in Amazon S3 where Synthetics stores artifacts from the
test runs of this canary. Artifacts include the log file, screenshots,
and HAR files.}

\item{ExecutionRoleArn}{[required] The ARN of the IAM role to be used to run the canary. This role must
already exist, and must include \code{lambda.amazonaws.com} as a principal in
the trust policy. The role must also have the following permissions:
\itemize{
\item \code{s3:PutObject}
\item \code{s3:GetBucketLocation}
\item \code{s3:ListAllMyBuckets}
\item \code{cloudwatch:PutMetricData}
\item \code{logs:CreateLogGroup}
\item \code{logs:CreateLogStream}
\item \code{logs:CreateLogStream}
}}

\item{Schedule}{[required] A structure that contains information about how often the canary is to
run and when these test runs are to stop.}

\item{RunConfig}{A structure that contains the configuration for individual canary runs,
such as timeout value.}

\item{SuccessRetentionPeriodInDays}{The number of days to retain data about successful runs of this canary.
If you omit this field, the default of 31 days is used. The valid range
is 1 to 455 days.}

\item{FailureRetentionPeriodInDays}{The number of days to retain data about failed runs of this canary. If
you omit this field, the default of 31 days is used. The valid range is
1 to 455 days.}

\item{RuntimeVersion}{[required] Specifies the runtime version to use for the canary. Currently, the only
valid value is \code{syn-1.0}. For more information about runtime versions,
see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_Library.html}{Canary Runtime Versions}.}

\item{VpcConfig}{If this canary is to test an endpoint in a VPC, this structure contains
information about the subnet and security groups of the VPC endpoint.
For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_VPC.html}{Running a Canary in a VPC}.}

\item{Tags}{A list of key-value pairs to associate with the canary. You can
associate as many as 50 tags with a canary.

Tags can help you organize and categorize your resources. You can also
use them to scope user permissions, by granting a user permission to
access or change only the resources that have certain tag values.}
}
\description{
Creates a canary. Canaries are scripts that monitor your endpoints and
APIs from the outside-in. Canaries help you check the availability and
latency of your web services and troubleshoot anomalies by investigating
load time data, screenshots of the UI, logs, and metrics. You can set up
a canary to run continuously or just once.
}
\details{
Do not use \code{CreateCanary} to modify an existing canary. Use
\href{https://docs.aws.amazon.com/AmazonSynthetics/latest/APIReference/API_UpdateCanary.html}{UpdateCanary}
instead.

To create canaries, you must have the \code{CloudWatchSyntheticsFullAccess}
policy. If you are creating a new IAM role for the canary, you also need
the the \code{iam:CreateRole}, \code{iam:CreatePolicy} and \code{iam:AttachRolePolicy}
permissions. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_Roles}{Necessary Roles and Permissions}.

Do not include secrets or proprietary information in your canary names.
The canary name makes up part of the Amazon Resource Name (ARN) for the
canary, and the ARN is included in outbound calls over the internet. For
more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/servicelens_canaries_security.html}{Security Considerations for Synthetics Canaries}.
}
\section{Request syntax}{
\preformatted{svc$create_canary(
  Name = "string",
  Code = list(
    S3Bucket = "string",
    S3Key = "string",
    S3Version = "string",
    ZipFile = raw,
    Handler = "string"
  ),
  ArtifactS3Location = "string",
  ExecutionRoleArn = "string",
  Schedule = list(
    Expression = "string",
    DurationInSeconds = 123
  ),
  RunConfig = list(
    TimeoutInSeconds = 123,
    MemoryInMB = 123
  ),
  SuccessRetentionPeriodInDays = 123,
  FailureRetentionPeriodInDays = 123,
  RuntimeVersion = "string",
  VpcConfig = list(
    SubnetIds = list(
      "string"
    ),
    SecurityGroupIds = list(
      "string"
    )
  ),
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
