% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_describe_canaries}
\alias{synthetics_describe_canaries}
\title{This operation returns a list of the canaries in your account, along
with full details about each canary}
\usage{
synthetics_describe_canaries(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{A token that indicates that there is more data available. You can use
this token in a subsequent operation to retrieve the next set of
results.}

\item{MaxResults}{Specify this parameter to limit how many canaries are returned each time
you use the \code{\link[=synthetics_describe_canaries]{describe_canaries}}
operation. If you omit this parameter, the default of 100 is used.}
}
\description{
This operation returns a list of the canaries in your account, along
with full details about each canary.

This operation does not have resource-level authorization, so if a user
is able to use \code{\link[=synthetics_describe_canaries]{describe_canaries}}, the
user can see all of the canaries in the account. A deny policy can only
be used to restrict access to all canaries. It cannot be used on
specific resources.
}
\section{Request syntax}{
\preformatted{svc$describe_canaries(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
