% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_describe_canaries_last_run}
\alias{synthetics_describe_canaries_last_run}
\title{Use this operation to see information from the most recent run of each
canary that you have created}
\usage{
synthetics_describe_canaries_last_run(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{A token that indicates that there is more data available. You can use
this token in a subsequent \code{DescribeCanaries} operation to retrieve the
next set of results.}

\item{MaxResults}{Specify this parameter to limit how many runs are returned each time you
use the \code{DescribeLastRun} operation. If you omit this parameter, the
default of 100 is used.}
}
\description{
Use this operation to see information from the most recent run of each
canary that you have created.
}
\section{Request syntax}{
\preformatted{svc$describe_canaries_last_run(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
