% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_describe_runtime_versions}
\alias{synthetics_describe_runtime_versions}
\title{Returns a list of Synthetics canary runtime versions}
\usage{
synthetics_describe_runtime_versions(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{A token that indicates that there is more data available. You can use
this token in a subsequent
\code{\link[=synthetics_describe_runtime_versions]{describe_runtime_versions}}
operation to retrieve the next set of results.}

\item{MaxResults}{Specify this parameter to limit how many runs are returned each time you
use the
\code{\link[=synthetics_describe_runtime_versions]{describe_runtime_versions}}
operation. If you omit this parameter, the default of 100 is used.}
}
\description{
Returns a list of Synthetics canary runtime versions. For more
information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_Library.html}{Canary Runtime Versions}.
}
\section{Request syntax}{
\preformatted{svc$describe_runtime_versions(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
