% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_get_canary_runs}
\alias{synthetics_get_canary_runs}
\title{Retrieves a list of runs for a specified canary}
\usage{
synthetics_get_canary_runs(Name, NextToken, MaxResults)
}
\arguments{
\item{Name}{[required] The name of the canary that you want to see runs for.}

\item{NextToken}{A token that indicates that there is more data available. You can use
this token in a subsequent \code{GetCanaryRuns} operation to retrieve the
next set of results.}

\item{MaxResults}{Specify this parameter to limit how many runs are returned each time you
use the \code{GetCanaryRuns} operation. If you omit this parameter, the
default of 100 is used.}
}
\description{
Retrieves a list of runs for a specified canary.
}
\section{Request syntax}{
\preformatted{svc$get_canary_runs(
  Name = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
