% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_list_tags_for_resource}
\alias{synthetics_list_tags_for_resource}
\title{Displays the tags associated with a canary}
\usage{
synthetics_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the canary that you want to view tags for.

The ARN format of a canary is
\verb{arn:aws:synthetics:<i>Region</i>:<i>account-id</i>:canary:<i>canary-name</i> }.}
}
\description{
Displays the tags associated with a canary.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
