% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_start_canary}
\alias{synthetics_start_canary}
\title{Use this operation to run a canary that has already been created}
\usage{
synthetics_start_canary(Name)
}
\arguments{
\item{Name}{[required] The name of the canary that you want to run. To find canary names, use
\href{https://docs.aws.amazon.com/AmazonSynthetics/latest/APIReference/API_DescribeCanaries.html}{DescribeCanaries}.}
}
\description{
Use this operation to run a canary that has already been created. The
frequency of the canary runs is determined by the value of the canary's
\code{Schedule}. To see a canary's schedule, use
\href{https://docs.aws.amazon.com/AmazonSynthetics/latest/APIReference/API_GetCanary.html}{GetCanary}.
}
\section{Request syntax}{
\preformatted{svc$start_canary(
  Name = "string"
)
}
}

\keyword{internal}
