% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_tag_resource}
\alias{synthetics_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified canary}
\usage{
synthetics_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the canary that you're adding tags to.

The ARN format of a canary is
\verb{arn:aws:synthetics:<i>Region</i>:<i>account-id</i>:canary:<i>canary-name</i> }.}

\item{Tags}{[required] The list of key-value pairs to associate with the canary.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified canary.

Tags can help you organize and categorize your resources. You can also
use them to scope user permissions, by granting a user permission to
access or change only resources with certain tag values.

Tags don't have any semantic meaning to AWS and are interpreted strictly
as strings of characters.

You can use the \code{TagResource} action with a canary that already has
tags. If you specify a new tag key for the alarm, this tag is appended
to the list of tags associated with the alarm. If you specify a tag key
that is already associated with the alarm, the new tag value that you
specify replaces the previous value for that tag.

You can associate as many as 50 tags with a canary.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
