% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_untag_resource}
\alias{synthetics_untag_resource}
\title{Removes one or more tags from the specified canary}
\usage{
synthetics_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the canary that you're removing tags from.

The ARN format of a canary is
\verb{arn:aws:synthetics:<i>Region</i>:<i>account-id</i>:canary:<i>canary-name</i> }.}

\item{TagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified canary.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
