% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_update_canary}
\alias{synthetics_update_canary}
\title{Use this operation to change the settings of a canary that has already
been created}
\usage{
synthetics_update_canary(Name, Code, ExecutionRoleArn, RuntimeVersion,
  Schedule, RunConfig, SuccessRetentionPeriodInDays,
  FailureRetentionPeriodInDays, VpcConfig)
}
\arguments{
\item{Name}{[required] The name of the canary that you want to update. To find the names of
your canaries, use
\href{https://docs.aws.amazon.com/AmazonSynthetics/latest/APIReference/API_DescribeCanaries.html}{DescribeCanaries}.

You cannot change the name of a canary that has already been created.}

\item{Code}{A structure that includes the entry point from which the canary should
start running your script. If the script is stored in an S3 bucket, the
bucket name, key, and version are also included.}

\item{ExecutionRoleArn}{The ARN of the IAM role to be used to run the canary. This role must
already exist, and must include \code{lambda.amazonaws.com} as a principal in
the trust policy. The role must also have the following permissions:
\itemize{
\item \code{s3:PutObject}
\item \code{s3:GetBucketLocation}
\item \code{s3:ListAllMyBuckets}
\item \code{cloudwatch:PutMetricData}
\item \code{logs:CreateLogGroup}
\item \code{logs:CreateLogStream}
\item \code{logs:CreateLogStream}
}}

\item{RuntimeVersion}{Specifies the runtime version to use for the canary. For a list of valid
runtime versions and for more information about runtime versions, see
\href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_Library.html}{Canary Runtime Versions}.}

\item{Schedule}{A structure that contains information about how often the canary is to
run, and when these runs are to stop.}

\item{RunConfig}{A structure that contains the timeout value that is used for each
individual run of the canary.}

\item{SuccessRetentionPeriodInDays}{The number of days to retain data about successful runs of this canary.}

\item{FailureRetentionPeriodInDays}{The number of days to retain data about failed runs of this canary.}

\item{VpcConfig}{If this canary is to test an endpoint in a VPC, this structure contains
information about the subnet and security groups of the VPC endpoint.
For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_VPC.html}{Running a Canary in a VPC}.}
}
\description{
Use this operation to change the settings of a canary that has already
been created.
}
\details{
You can't use this operation to update the tags of an existing canary.
To change the tags of an existing canary, use
\href{https://docs.aws.amazon.com/AmazonSynthetics/latest/APIReference/API_TagResource.html}{TagResource}.
}
\section{Request syntax}{
\preformatted{svc$update_canary(
  Name = "string",
  Code = list(
    S3Bucket = "string",
    S3Key = "string",
    S3Version = "string",
    ZipFile = raw,
    Handler = "string"
  ),
  ExecutionRoleArn = "string",
  RuntimeVersion = "string",
  Schedule = list(
    Expression = "string",
    DurationInSeconds = 123
  ),
  RunConfig = list(
    TimeoutInSeconds = 123,
    MemoryInMB = 123,
    ActiveTracing = TRUE|FALSE,
    EnvironmentVariables = list(
      "string"
    )
  ),
  SuccessRetentionPeriodInDays = 123,
  FailureRetentionPeriodInDays = 123,
  VpcConfig = list(
    SubnetIds = list(
      "string"
    ),
    SecurityGroupIds = list(
      "string"
    )
  )
)
}
}

\keyword{internal}
