% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_cancel_query}
\alias{timestreamquery_cancel_query}
\title{Cancels a query that has been issued}
\usage{
timestreamquery_cancel_query(QueryId)
}
\arguments{
\item{QueryId}{[required] The id of the query that needs to be cancelled. \code{QueryID} is returned as
part of QueryResult.}
}
\description{
Cancels a query that has been issued. Cancellation is guaranteed only if
the query has not completed execution before the cancellation request
was issued. Because cancellation is an idempotent operation, subsequent
cancellation requests will return a \code{CancellationMessage}, indicating
that the query has already been canceled.
}
\section{Request syntax}{
\preformatted{svc$cancel_query(
  QueryId = "string"
)
}
}

\keyword{internal}
