% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_query}
\alias{timestreamquery_query}
\title{Query is a synchronous operation that enables you to execute a query}
\usage{
timestreamquery_query(QueryString, ClientToken, NextToken, MaxRows)
}
\arguments{
\item{QueryString}{[required] The query to be executed by Timestream.}

\item{ClientToken}{Unique, case-sensitive string of up to 64 ASCII characters that you
specify when you make a Query request. Providing a \code{ClientToken} makes
the call to \code{Query} idempotent, meaning that multiple identical calls
have the same effect as one single call.

Your query request will fail in the following cases:
\itemize{
\item If you submit a request with the same client token outside the
5-minute idepotency window.
\item If you submit a request with the same client token but a change in
other parameters within the 5-minute idempotency window.
}

After 4 hours, any request with the same client token is treated as a
new request.}

\item{NextToken}{A pagination token passed to get a set of results.}

\item{MaxRows}{The total number of rows to return in the output. If the total number of
rows available is more than the value specified, a NextToken is provided
in the command's output. To resume pagination, provide the NextToken
value in the starting-token argument of a subsequent command.}
}
\description{
Query is a synchronous operation that enables you to execute a query.
Query will timeout after 60 seconds. You must update the default timeout
in the SDK to support a timeout of 60 seconds. The result set will be
truncated to 1MB. Service quotas apply. For more information, see Quotas
in the Timestream Developer Guide.
}
\section{Request syntax}{
\preformatted{svc$query(
  QueryString = "string",
  ClientToken = "string",
  NextToken = "string",
  MaxRows = 123
)
}
}

\keyword{internal}
