% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_create_database}
\alias{timestreamwrite_create_database}
\title{Creates a new Timestream database}
\usage{
timestreamwrite_create_database(DatabaseName, KmsKeyId, Tags)
}
\arguments{
\item{DatabaseName}{[required] The name of the Timestream database.}

\item{KmsKeyId}{The KMS key for the database. If the KMS key is not specified, the
database will be encrypted with a Timestream managed KMS key located in
your account. Refer to \href{https://docs.aws.amazon.com/kms/latest/developerguide/concepts.html#aws-managed-cmk}{AWS managed KMS keys}
for more info.}

\item{Tags}{A list of key-value pairs to label the table.}
}
\description{
Creates a new Timestream database. If the KMS key is not specified, the
database will be encrypted with a Timestream managed KMS key located in
your account. Refer to \href{https://docs.aws.amazon.com/kms/latest/developerguide/concepts.html#aws-managed-cmk}{AWS managed KMS keys}
for more info. Service quotas apply. For more information, see \href{https://docs.aws.amazon.com/timestream/latest/developerguide/ts-limits.html}{Access Management}
in the Timestream Developer Guide.
}
\section{Request syntax}{
\preformatted{svc$create_database(
  DatabaseName = "string",
  KmsKeyId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
