% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_create_table}
\alias{timestreamwrite_create_table}
\title{The CreateTable operation adds a new table to an existing database in
your account}
\usage{
timestreamwrite_create_table(DatabaseName, TableName,
  RetentionProperties, Tags)
}
\arguments{
\item{DatabaseName}{[required] The name of the Timestream database.}

\item{TableName}{[required] The name of the Timestream table.}

\item{RetentionProperties}{The duration for which your time series data must be stored in the
memory store and the magnetic store.}

\item{Tags}{A list of key-value pairs to label the table.}
}
\description{
The CreateTable operation adds a new table to an existing database in
your account. In an AWS account, table names must be at least unique
within each Region if they are in the same database. You may have
identical table names in the same Region if the tables are in seperate
databases. While creating the table, you must specify the table name,
database name, and the retention properties. Service quotas apply. For
more information, see \href{https://docs.aws.amazon.com/timestream/latest/developerguide/ts-limits.html}{Access Management}
in the Timestream Developer Guide.
}
\section{Request syntax}{
\preformatted{svc$create_table(
  DatabaseName = "string",
  TableName = "string",
  RetentionProperties = list(
    MemoryStoreRetentionPeriodInHours = 123,
    MagneticStoreRetentionPeriodInDays = 123
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
