% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_delete_database}
\alias{timestreamwrite_delete_database}
\title{Deletes a given Timestream database}
\usage{
timestreamwrite_delete_database(DatabaseName)
}
\arguments{
\item{DatabaseName}{[required] The name of the Timestream database to be deleted.}
}
\description{
Deletes a given Timestream database. \emph{This is an irreversible operation.
After a database is deleted, the time series data from its tables cannot
be recovered.}
}
\details{
All tables in the database must be deleted first, or a
ValidationException error will be thrown.

Due to the nature of distributed retries, the operation can return
either success or a ResourceNotFoundException. Clients should consider
them equivalent.
}
\section{Request syntax}{
\preformatted{svc$delete_database(
  DatabaseName = "string"
)
}
}

\keyword{internal}
