% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_delete_table}
\alias{timestreamwrite_delete_table}
\title{Deletes a given Timestream table}
\usage{
timestreamwrite_delete_table(DatabaseName, TableName)
}
\arguments{
\item{DatabaseName}{[required] The name of the database where the Timestream database is to be deleted.}

\item{TableName}{[required] The name of the Timestream table to be deleted.}
}
\description{
Deletes a given Timestream table. This is an irreversible operation.
After a Timestream database table is deleted, the time series data
stored in the table cannot be recovered.
}
\details{
Due to the nature of distributed retries, the operation can return
either success or a ResourceNotFoundException. Clients should consider
them equivalent.
}
\section{Request syntax}{
\preformatted{svc$delete_table(
  DatabaseName = "string",
  TableName = "string"
)
}
}

\keyword{internal}
