% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_describe_endpoints}
\alias{timestreamwrite_describe_endpoints}
\title{DescribeEndpoints returns a list of available endpoints to make
Timestream API calls against}
\usage{
timestreamwrite_describe_endpoints()
}
\description{
DescribeEndpoints returns a list of available endpoints to make
Timestream API calls against. This API is available through both Write
and Query.
}
\details{
Because Timestream’s SDKs are designed to transparently work with the
service’s architecture, including the management and mapping of the
service endpoints, \emph{it is not recommended that you use this API unless}:
\itemize{
\item Your application uses a programming language that does not yet have
SDK support
\item You require better control over the client-side implementation
}

For detailed information on how to use DescribeEndpoints, see \href{https://docs.aws.amazon.com/timestream/latest/developerguide/Using-API.endpoint-discovery.html}{The Endpoint Discovery Pattern and REST APIs}.
}
\section{Request syntax}{
\preformatted{svc$describe_endpoints()
}
}

\keyword{internal}
