% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_describe_table}
\alias{timestreamwrite_describe_table}
\title{Returns information about the table, including the table name, database
name, retention duration of the memory store and the magnetic store}
\usage{
timestreamwrite_describe_table(DatabaseName, TableName)
}
\arguments{
\item{DatabaseName}{[required] The name of the Timestream database.}

\item{TableName}{[required] The name of the Timestream table.}
}
\description{
Returns information about the table, including the table name, database
name, retention duration of the memory store and the magnetic store.
Service quotas apply. For more information, see \href{https://docs.aws.amazon.com/timestream/latest/developerguide/ts-limits.html}{Access Management}
in the Timestream Developer Guide.
}
\section{Request syntax}{
\preformatted{svc$describe_table(
  DatabaseName = "string",
  TableName = "string"
)
}
}

\keyword{internal}
