% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_list_tags_for_resource}
\alias{timestreamwrite_list_tags_for_resource}
\title{List all tags on a Timestream resource}
\usage{
timestreamwrite_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Timestream resource with tags to be listed. This value is an Amazon
Resource Name (ARN).}
}
\description{
List all tags on a Timestream resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}
}

\keyword{internal}
