% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_medical_transcription_job}
\alias{transcribeservice_delete_medical_transcription_job}
\title{Deletes a transcription job generated by Amazon Transcribe Medical and
any related information}
\usage{
transcribeservice_delete_medical_transcription_job(
  MedicalTranscriptionJobName)
}
\arguments{
\item{MedicalTranscriptionJobName}{[required] The name you provide to the
\code{\link[=transcribeservice_delete_medical_transcription_job]{delete_medical_transcription_job}}
object to delete a transcription job.}
}
\description{
Deletes a transcription job generated by Amazon Transcribe Medical and
any related information.
}
\section{Request syntax}{
\preformatted{svc$delete_medical_transcription_job(
  MedicalTranscriptionJobName = "string"
)
}
}

\keyword{internal}
