% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_medical_vocabulary}
\alias{transcribeservice_delete_medical_vocabulary}
\title{Deletes a vocabulary from Amazon Transcribe Medical}
\usage{
transcribeservice_delete_medical_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{[required] The name of the vocabulary you are choosing to delete.}
}
\description{
Deletes a vocabulary from Amazon Transcribe Medical.
}
\section{Request syntax}{
\preformatted{svc$delete_medical_vocabulary(
  VocabularyName = "string"
)
}
}

\keyword{internal}
