% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_medical_vocabulary}
\alias{transcribeservice_get_medical_vocabulary}
\title{Retrieves information about a medical vocabulary}
\usage{
transcribeservice_get_medical_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{[required] The name of the vocabulary that you want information about. The value is
case sensitive.}
}
\description{
Retrieves information about a medical vocabulary.
}
\section{Request syntax}{
\preformatted{svc$get_medical_vocabulary(
  VocabularyName = "string"
)
}
}

\keyword{internal}
