% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_transcription_job}
\alias{transcribeservice_get_transcription_job}
\title{Returns information about a transcription job}
\usage{
transcribeservice_get_transcription_job(TranscriptionJobName)
}
\arguments{
\item{TranscriptionJobName}{[required] The name of the job.}
}
\description{
Returns information about a transcription job. To see the status of the
job, check the \code{TranscriptionJobStatus} field. If the status is
\code{COMPLETED}, the job is finished and you can find the results at the
location specified in the \code{TranscriptFileUri} field. If you enable
content redaction, the redacted transcript appears in
\code{RedactedTranscriptFileUri}.
}
\section{Request syntax}{
\preformatted{svc$get_transcription_job(
  TranscriptionJobName = "string"
)
}
}

\keyword{internal}
