% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_vocabulary}
\alias{transcribeservice_get_vocabulary}
\title{Gets information about a vocabulary}
\usage{
transcribeservice_get_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{[required] The name of the vocabulary to return information about. The name is case
sensitive.}
}
\description{
Gets information about a vocabulary.
}
\section{Request syntax}{
\preformatted{svc$get_vocabulary(
  VocabularyName = "string"
)
}
}

\keyword{internal}
