% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_list_vocabulary_filters}
\alias{transcribeservice_list_vocabulary_filters}
\title{Gets information about vocabulary filters}
\usage{
transcribeservice_list_vocabulary_filters(NextToken, MaxResults,
  NameContains)
}
\arguments{
\item{NextToken}{If the result of the previous request to \code{ListVocabularyFilters} was
truncated, include the \code{NextToken} to fetch the next set of collections.}

\item{MaxResults}{The maximum number of filters to return in the response. If there are
fewer results in the list, this response contains only the actual
results.}

\item{NameContains}{Filters the response so that it only contains vocabulary filters whose
name contains the specified string.}
}
\description{
Gets information about vocabulary filters.
}
\section{Request syntax}{
\preformatted{svc$list_vocabulary_filters(
  NextToken = "string",
  MaxResults = 123,
  NameContains = "string"
)
}
}

\keyword{internal}
