% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_update_vocabulary}
\alias{transcribeservice_update_vocabulary}
\title{Updates an existing vocabulary with new values}
\usage{
transcribeservice_update_vocabulary(VocabularyName, LanguageCode,
  Phrases)
}
\arguments{
\item{VocabularyName}{[required] The name of the vocabulary to update. The name is case-sensitive.}

\item{LanguageCode}{[required] The language code of the vocabulary entries.}

\item{Phrases}{[required] An array of strings containing the vocabulary entries.}
}
\description{
Updates an existing vocabulary with new values. The \code{UpdateVocabulary}
operation overwrites all of the existing information with the values
that you provide in the request.
}
\section{Request syntax}{
\preformatted{transcribeservice$update_vocabulary(
  VocabularyName = "string",
  LanguageCode = "en-US"|"es-US"|"en-AU"|"fr-CA"|"en-GB"|"de-DE"|"pt-BR"|"fr-FR"|"it-IT",
  Phrases = list(
    "string"
  )
)
}
}

\keyword{internal}
