% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_delete_ssh_public_key}
\alias{transfer_delete_ssh_public_key}
\title{Deletes a user's Secure Shell (SSH) public key}
\usage{
transfer_delete_ssh_public_key(ServerId, SshPublicKeyId, UserName)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for a file transfer protocol-enabled
server instance that has the user assigned to it.}

\item{SshPublicKeyId}{[required] A unique identifier used to reference your user's specific SSH key.}

\item{UserName}{[required] A unique string that identifies a user whose public key is being
deleted.}
}
\description{
Deletes a user's Secure Shell (SSH) public key.
}
\details{
No response is returned from this operation.
}
\section{Request syntax}{
\preformatted{svc$delete_ssh_public_key(
  ServerId = "string",
  SshPublicKeyId = "string",
  UserName = "string"
)
}
}

\keyword{internal}
