% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_list_users}
\alias{transfer_list_users}
\title{Lists the users for the server that you specify by passing the ServerId
parameter}
\usage{
transfer_list_users(MaxResults, NextToken, ServerId)
}
\arguments{
\item{MaxResults}{Specifies the number of users to return as a response to the \code{ListUsers}
request.}

\item{NextToken}{When you can get additional results from the \code{ListUsers} call, a
\code{NextToken} parameter is returned in the output. You can then pass in a
subsequent command to the \code{NextToken} parameter to continue listing
additional users.}

\item{ServerId}{[required] A system-assigned unique identifier for a Secure File Transfer Protocol
(SFTP) server that has users assigned to it.}
}
\description{
Lists the users for the server that you specify by passing the
\code{ServerId} parameter.
}
\section{Request syntax}{
\preformatted{svc$list_users(
  MaxResults = 123,
  NextToken = "string",
  ServerId = "string"
)
}
}

\keyword{internal}
