% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_stop_server}
\alias{transfer_stop_server}
\title{Changes the state of a file transfer protocol-enabled server from ONLINE
to OFFLINE}
\usage{
transfer_stop_server(ServerId)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for a file transfer protocol-enabled
server that you stopped.}
}
\description{
Changes the state of a file transfer protocol-enabled server from
\code{ONLINE} to \code{OFFLINE}. An \code{OFFLINE} server cannot accept and process
file transfer jobs. Information tied to your server, such as server and
user properties, are not affected by stopping your server. Stopping the
server will not reduce or impact your file transfer protocol endpoint
billing.
}
\details{
The state of \code{STOPPING} indicates that the server is in an intermediate
state, either not fully able to respond, or not fully offline. The
values of \code{STOP_FAILED} can indicate an error condition.

No response is returned from this call.
}
\section{Request syntax}{
\preformatted{svc$stop_server(
  ServerId = "string"
)
}
}

\keyword{internal}
