% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_tag_resource}
\alias{transfer_tag_resource}
\title{Attaches a key-value pair to a resource, as identified by its Amazon
Resource Name (ARN)}
\usage{
transfer_tag_resource(Arn, Tags)
}
\arguments{
\item{Arn}{[required] An Amazon Resource Name (ARN) for a specific AWS resource, such as a
server, user, or role.}

\item{Tags}{[required] Key-value pairs assigned to ARNs that you can use to group and search
for resources by type. You can attach this metadata to user accounts for
any purpose.}
}
\description{
Attaches a key-value pair to a resource, as identified by its Amazon
Resource Name (ARN). Resources are users, servers, roles, and other
entities.
}
\details{
There is no response returned from this call.
}
\section{Request syntax}{
\preformatted{transfer$tag_resource(
  Arn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
