% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_update_user}
\alias{transfer_update_user}
\title{Assigns new properties to a user}
\usage{
transfer_update_user(HomeDirectory, Policy, Role, ServerId, UserName)
}
\arguments{
\item{HomeDirectory}{The HomeDirectory parameter specifies the landing directory (folder) for
a user when they log in to the server using their client. An example
would be: \code{/home/<i>username</i> }.}

\item{Policy}{Allows you to supply a scope-down policy for your user so you can use
the same AWS Identity and Access Management (IAM) role across multiple
users. The policy scopes down users access to portions of your Amazon S3
bucket. Variables you can use inside this policy include
\code{$\{Transfer:UserName\}}, \code{$\{Transfer:HomeDirectory\}}, and
\code{$\{Transfer:HomeBucket\}}.}

\item{Role}{The IAM role that controls your user's access to your Amazon S3 bucket.
The policies attached to this role will determine the level of access
you want to provide your users when transferring files into and out of
your Amazon S3 bucket or buckets. The IAM role should also contain a
trust relationship that allows the Secure File Transfer Protocol (SFTP)
server to access your resources when servicing your SFTP user's transfer
requests.}

\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server instance that the
user account is assigned to.}

\item{UserName}{[required] A unique string that identifies a user and is associated with a server
as specified by the ServerId. This is the string that will be used by
your user when they log in to your SFTP server.}
}
\description{
Assigns new properties to a user. Parameters you pass modify any or all
of the following: the home directory, role, and policy for the
\code{UserName} and \code{ServerId} you specify.
}
\details{
The response returns the \code{ServerId} and the \code{UserName} for the updated
user.
}
\section{Request syntax}{
\preformatted{transfer$update_user(
  HomeDirectory = "string",
  Policy = "string",
  Role = "string",
  ServerId = "string",
  UserName = "string"
)
}
}

\keyword{internal}
