% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_create_rule_group}
\alias{waf_create_rule_group}
\title{This is AWS WAF Classic documentation}
\usage{
waf_create_rule_group(Name, MetricName, ChangeToken, Tags)
}
\arguments{
\item{Name}{[required] A friendly name or description of the RuleGroup. You can't change \code{Name}
after you create a \code{RuleGroup}.}

\item{MetricName}{[required] A friendly name or description for the metrics for this \code{RuleGroup}. The
name can contain only alphanumeric characters (A-Z, a-z, 0-9), with
maximum length 128 and minimum length one. It can't contain whitespace
or metric names reserved for AWS WAF, including "All" and
"Default\\_Action." You can't change the name of the metric after you
create the \code{RuleGroup}.}

\item{ChangeToken}{[required] The value returned by the most recent call to GetChangeToken.}

\item{Tags}{}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.
}
\details{
\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Creates a \code{RuleGroup}. A rule group is a collection of predefined rules
that you add to a web ACL. You use UpdateRuleGroup to add rules to the
rule group.

Rule groups are subject to the following limits:
\itemize{
\item Three rule groups per account. You can request an increase to this
limit by contacting customer support.
\item One rule group per web ACL.
\item Ten rules per rule group.
}

For more information about how to use the AWS WAF API to allow or block
HTTP requests, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/}{AWS WAF Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_rule_group(
  Name = "string",
  MetricName = "string",
  ChangeToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
