% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_delete_regex_pattern_set}
\alias{waf_delete_regex_pattern_set}
\title{Permanently deletes a RegexPatternSet}
\usage{
waf_delete_regex_pattern_set(RegexPatternSetId, ChangeToken)
}
\arguments{
\item{RegexPatternSetId}{[required] The \code{RegexPatternSetId} of the RegexPatternSet that you want to delete.
\code{RegexPatternSetId} is returned by CreateRegexPatternSet and by
ListRegexPatternSets.}

\item{ChangeToken}{[required] The value returned by the most recent call to GetChangeToken.}
}
\description{
Permanently deletes a RegexPatternSet. You can't delete a
\code{RegexPatternSet} if it's still used in any \code{RegexMatchSet} or if the
\code{RegexPatternSet} is not empty.
}
\section{Request syntax}{
\preformatted{waf$delete_regex_pattern_set(
  RegexPatternSetId = "string",
  ChangeToken = "string"
)
}
}

\keyword{internal}
