% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_delete_rule}
\alias{waf_delete_rule}
\title{This is AWS WAF Classic documentation}
\usage{
waf_delete_rule(RuleId, ChangeToken)
}
\arguments{
\item{RuleId}{[required] The \code{RuleId} of the Rule that you want to delete. \code{RuleId} is returned
by \code{\link[=waf_create_rule]{create_rule}} and by
\code{\link[=waf_list_rules]{list_rules}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=waf_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Permanently deletes a Rule. You can't delete a \code{Rule} if it's still used
in any \code{WebACL} objects or if it still includes any predicates, such as
\code{ByteMatchSet} objects.

If you just want to remove a \code{Rule} from a \code{WebACL}, use
\code{\link[=waf_update_web_acl]{update_web_acl}}.

To permanently delete a \code{Rule} from AWS WAF, perform the following
steps:
\enumerate{
\item Update the \code{Rule} to remove predicates, if any. For more
information, see \code{\link[=waf_update_rule]{update_rule}}.
\item Use \code{\link[=waf_get_change_token]{get_change_token}} to get the change
token that you provide in the \code{ChangeToken} parameter of a
\code{\link[=waf_delete_rule]{delete_rule}} request.
\item Submit a \code{\link[=waf_delete_rule]{delete_rule}} request.
}
}
\section{Request syntax}{
\preformatted{svc$delete_rule(
  RuleId = "string",
  ChangeToken = "string"
)
}
}

\examples{
\dontrun{
# The following example deletes a rule with the ID WAFRule-1-Example.
svc$delete_rule(
  ChangeToken = "abcd12f2-46da-4fdb-b8d5-fbd4c466928f",
  RuleId = "WAFRule-1-Example"
)
}

}
\keyword{internal}
