% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_logging_configuration}
\alias{waf_get_logging_configuration}
\title{Returns the LoggingConfiguration for the specified web ACL}
\usage{
waf_get_logging_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the web ACL for which you want to get
the LoggingConfiguration.}
}
\description{
Returns the LoggingConfiguration for the specified web ACL.
}
\section{Request syntax}{
\preformatted{svc$get_logging_configuration(
  ResourceArn = "string"
)
}
}

\keyword{internal}
