% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_delete_sql_injection_match_set}
\alias{wafregional_delete_sql_injection_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_delete_sql_injection_match_set(SqlInjectionMatchSetId,
  ChangeToken)
}
\arguments{
\item{SqlInjectionMatchSetId}{[required] The \code{SqlInjectionMatchSetId} of the SqlInjectionMatchSet that you want
to delete. \code{SqlInjectionMatchSetId} is returned by
\code{\link[=wafregional_create_sql_injection_match_set]{create_sql_injection_match_set}}
and by
\code{\link[=wafregional_list_sql_injection_match_sets]{list_sql_injection_match_sets}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Permanently deletes a SqlInjectionMatchSet. You can't delete a
\code{SqlInjectionMatchSet} if it's still used in any \code{Rules} or if it still
contains any SqlInjectionMatchTuple objects.

If you just want to remove a \code{SqlInjectionMatchSet} from a \code{Rule}, use
\code{\link[=wafregional_update_rule]{update_rule}}.

To permanently delete a \code{SqlInjectionMatchSet} from AWS WAF, perform the
following steps:
\enumerate{
\item Update the \code{SqlInjectionMatchSet} to remove filters, if any. For
more information, see
\code{\link[=wafregional_update_sql_injection_match_set]{update_sql_injection_match_set}}.
\item Use \code{\link[=wafregional_get_change_token]{get_change_token}} to get the
change token that you provide in the \code{ChangeToken} parameter of a
\code{\link[=wafregional_delete_sql_injection_match_set]{delete_sql_injection_match_set}}
request.
\item Submit a
\code{\link[=wafregional_delete_sql_injection_match_set]{delete_sql_injection_match_set}}
request.
}
}
\section{Request syntax}{
\preformatted{svc$delete_sql_injection_match_set(
  SqlInjectionMatchSetId = "string",
  ChangeToken = "string"
)
}
}

\examples{
\dontrun{
# The following example deletes a SQL injection match set  with the ID
# example1ds3t-46da-4fdb-b8d5-abc321j569j5.
svc$delete_sql_injection_match_set(
  ChangeToken = "abcd12f2-46da-4fdb-b8d5-fbd4c466928f",
  SqlInjectionMatchSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)
}

}
\keyword{internal}
