% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_geo_match_set}
\alias{wafregional_get_geo_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_get_geo_match_set(GeoMatchSetId)
}
\arguments{
\item{GeoMatchSetId}{[required] The \code{GeoMatchSetId} of the GeoMatchSet that you want to get.
\code{GeoMatchSetId} is returned by
\code{\link[=wafregional_create_geo_match_set]{create_geo_match_set}} and by
\code{\link[=wafregional_list_geo_match_sets]{list_geo_match_sets}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Returns the GeoMatchSet that is specified by \code{GeoMatchSetId}.
}
\section{Request syntax}{
\preformatted{svc$get_geo_match_set(
  GeoMatchSetId = "string"
)
}
}

\keyword{internal}
