% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_sampled_requests}
\alias{wafregional_get_sampled_requests}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_get_sampled_requests(WebAclId, RuleId, TimeWindow, MaxItems)
}
\arguments{
\item{WebAclId}{[required] The \code{WebACLId} of the \code{WebACL} for which you want
\code{\link[=wafregional_get_sampled_requests]{get_sampled_requests}} to return a
sample of requests.}

\item{RuleId}{[required] \code{RuleId} is one of three values:
\itemize{
\item The \code{RuleId} of the \code{Rule} or the \code{RuleGroupId} of the \code{RuleGroup}
for which you want
\code{\link[=wafregional_get_sampled_requests]{get_sampled_requests}} to return
a sample of requests.
\item \code{Default_Action}, which causes
\code{\link[=wafregional_get_sampled_requests]{get_sampled_requests}} to return
a sample of the requests that didn't match any of the rules in the
specified \code{WebACL}.
}}

\item{TimeWindow}{[required] The start date and time and the end date and time of the range for which
you want \code{\link[=wafregional_get_sampled_requests]{get_sampled_requests}} to
return a sample of requests. You must specify the times in Coordinated
Universal Time (UTC) format. UTC format includes the special designator,
\code{Z}. For example, \code{"2016-09-27T14:50Z"}. You can specify any time range
in the previous three hours.}

\item{MaxItems}{[required] The number of requests that you want AWS WAF to return from among the
first 5,000 requests that your AWS resource received during the time
range. If your resource received fewer requests than the value of
\code{MaxItems}, \code{\link[=wafregional_get_sampled_requests]{get_sampled_requests}}
returns information about all of them.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Gets detailed information about a specified number of requests--a
sample--that AWS WAF randomly selects from among the first 5,000
requests that your AWS resource received during a time range that you
choose. You can specify a sample size of up to 500 requests, and you can
specify any time range in the previous three hours.

\code{\link[=wafregional_get_sampled_requests]{get_sampled_requests}} returns a
time range, which is usually the time range that you specified. However,
if your resource (such as a CloudFront distribution) received 5,000
requests before the specified time range elapsed,
\code{\link[=wafregional_get_sampled_requests]{get_sampled_requests}} returns an
updated time range. This new time range indicates the actual period
during which AWS WAF selected the requests in the sample.
}
\section{Request syntax}{
\preformatted{svc$get_sampled_requests(
  WebAclId = "string",
  RuleId = "string",
  TimeWindow = list(
    StartTime = as.POSIXct(
      "2015-01-01"
    ),
    EndTime = as.POSIXct(
      "2015-01-01"
    )
  ),
  MaxItems = 123
)
}
}

\examples{
\dontrun{
# The following example returns detailed information about 100 requests
# --a sample-- that AWS WAF randomly selects from among the first 5,000
# requests that your AWS resource received between the time period
# 2016-09-27T15:50Z to 2016-09-27T15:50Z.
svc$get_sampled_requests(
  MaxItems = 100L,
  RuleId = "WAFRule-1-Example",
  TimeWindow = list(
    EndTime = "2016-09-27T15:50Z",
    StartTime = "2016-09-27T15:50Z"
  ),
  WebAclId = "createwebacl-1472061481310"
)
}

}
\keyword{internal}
