% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_sql_injection_match_set}
\alias{wafregional_get_sql_injection_match_set}
\title{Returns the SqlInjectionMatchSet that is specified by
SqlInjectionMatchSetId}
\usage{
wafregional_get_sql_injection_match_set(SqlInjectionMatchSetId)
}
\arguments{
\item{SqlInjectionMatchSetId}{[required] The \code{SqlInjectionMatchSetId} of the SqlInjectionMatchSet that you want
to get. \code{SqlInjectionMatchSetId} is returned by
CreateSqlInjectionMatchSet and by ListSqlInjectionMatchSets.}
}
\description{
Returns the SqlInjectionMatchSet that is specified by
\code{SqlInjectionMatchSetId}.
}
\section{Request syntax}{
\preformatted{svc$get_sql_injection_match_set(
  SqlInjectionMatchSetId = "string"
)
}
}

\examples{
# The following example returns the details of a SQL injection match set
# with the ID example1ds3t-46da-4fdb-b8d5-abc321j569j5.
\donttest{svc$get_sql_injection_match_set(
  SqlInjectionMatchSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)}

}
\keyword{internal}
