% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_list_regex_pattern_sets}
\alias{wafregional_list_regex_pattern_sets}
\title{Returns an array of RegexPatternSetSummary objects}
\usage{
wafregional_list_regex_pattern_sets(NextMarker, Limit)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more \code{RegexPatternSet}
objects than the value of \code{Limit}, AWS WAF returns a \code{NextMarker} value
in the response that allows you to list another group of
\code{RegexPatternSet} objects. For the second and subsequent
\code{ListRegexPatternSets} requests, specify the value of \code{NextMarker} from
the previous response to get information about another batch of
\code{RegexPatternSet} objects.}

\item{Limit}{Specifies the number of \code{RegexPatternSet} objects that you want AWS WAF
to return for this request. If you have more \code{RegexPatternSet} objects
than the number you specify for \code{Limit}, the response includes a
\code{NextMarker} value that you can use to get another batch of
\code{RegexPatternSet} objects.}
}
\description{
Returns an array of RegexPatternSetSummary objects.
}
\section{Request syntax}{
\preformatted{wafregional$list_regex_pattern_sets(
  NextMarker = "string",
  Limit = 123
)
}
}

\keyword{internal}
