% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_tag_resource}
\alias{wafregional_tag_resource}
\title{Tag resource}
\usage{
wafregional_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required]}

\item{Tags}{[required]}
}
\description{
Tag resource
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceARN = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
