% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_associate_web_acl}
\alias{wafv2_associate_web_acl}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_associate_web_acl(WebACLArn, ResourceArn)
}
\arguments{
\item{WebACLArn}{[required] The Amazon Resource Name (ARN) of the Web ACL that you want to associate
with the resource.}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to associate with the web
ACL.

The ARN must be in one of the following formats:
\itemize{
\item For an Application Load Balancer:
\code{arn:aws:elasticloadbalancing:region:account-id:loadbalancer/app/load-balancer-name/load-balancer-id }
\item For an API Gateway REST API:
\verb{arn:aws:apigateway:region::/restapis/api-id/stages/stage-name }
\item For an AppSync GraphQL API:
\code{arn:aws:appsync:region:account-id:apis/GraphQLApiId }
}}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.

Associates a Web ACL with a regional application resource, to protect
the resource. A regional application can be an Application Load Balancer
(ALB), an API Gateway REST API, or an AppSync GraphQL API.

For AWS CloudFront, don't use this call. Instead, use your CloudFront
distribution configuration. To associate a Web ACL, in the CloudFront
call \code{UpdateDistribution}, set the web ACL ID to the Amazon Resource
Name (ARN) of the Web ACL. For information, see
\href{https://docs.aws.amazon.com/cloudfront/latest/APIReference/API_UpdateDistribution.html}{UpdateDistribution}.
}
\section{Request syntax}{
\preformatted{svc$associate_web_acl(
  WebACLArn = "string",
  ResourceArn = "string"
)
}
}

\keyword{internal}
