% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_create_regex_pattern_set}
\alias{wafv2_create_regex_pattern_set}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_create_regex_pattern_set(Name, Scope, Description,
  RegularExpressionList, Tags)
}
\arguments{
\item{Name}{[required] A friendly name of the set. You cannot change the name after you create
the set.}

\item{Scope}{[required] Specifies whether this is for an AWS CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB) or an API Gateway stage.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{Description}{A friendly description of the set. You cannot change the description of
a set after you create it.}

\item{RegularExpressionList}{[required] Array of regular expression strings.}

\item{Tags}{An array of key:value pairs to associate with the resource.}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
}
\details{
Creates a RegexPatternSet per the specifications provided.
}
\section{Request syntax}{
\preformatted{svc$create_regex_pattern_set(
  Name = "string",
  Scope = "CLOUDFRONT"|"REGIONAL",
  Description = "string",
  RegularExpressionList = list(
    list(
      RegexString = "string"
    )
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
