% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_delete_rule_group}
\alias{wafv2_delete_rule_group}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_delete_rule_group(Name, Scope, Id, LockToken)
}
\arguments{
\item{Name}{[required] The name of the rule group. You cannot change the name of a rule group
after you create it.}

\item{Scope}{[required] Specifies whether this is for an AWS CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{Id}{[required] A unique identifier for the rule group. This ID is returned in the
responses to create and list commands. You provide it to operations like
update and delete.}

\item{LockToken}{[required] A token used for optimistic locking. AWS WAF returns a token to your get
and list requests, to mark the state of the entity at the time of the
request. To make changes to the entity associated with the token, you
provide the token to operations like update and delete. AWS WAF uses the
token to ensure that no changes have been made to the entity since you
last retrieved it. If a change has been made, the update fails with a
\code{WAFOptimisticLockException}. If this happens, perform another get, and
use the new token returned by that operation.}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.

Deletes the specified RuleGroup.
}
\section{Request syntax}{
\preformatted{svc$delete_rule_group(
  Name = "string",
  Scope = "CLOUDFRONT"|"REGIONAL",
  Id = "string",
  LockToken = "string"
)
}
}

\keyword{internal}
