% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_describe_managed_rule_group}
\alias{wafv2_describe_managed_rule_group}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_describe_managed_rule_group(VendorName, Name, Scope)
}
\arguments{
\item{VendorName}{[required] The name of the managed rule group vendor. You use this, along with the
rule group name, to identify the rule group.}

\item{Name}{[required] The name of the managed rule group. You use this, along with the vendor
name, to identify the rule group.}

\item{Scope}{[required] Specifies whether this is for an AWS CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB) or an API Gateway stage.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
}
\details{
Provides high-level information for a managed rule group, including
descriptions of the rules.
}
\section{Request syntax}{
\preformatted{svc$describe_managed_rule_group(
  VendorName = "string",
  Name = "string",
  Scope = "CLOUDFRONT"|"REGIONAL"
)
}
}

\keyword{internal}
