% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_ip_set}
\alias{wafv2_get_ip_set}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_get_ip_set(Name, Scope, Id)
}
\arguments{
\item{Name}{[required] The name of the IP set. You cannot change the name of an \code{IPSet} after
you create it.}

\item{Scope}{[required] Specifies whether this is for an AWS CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{Id}{[required] A unique identifier for the set. This ID is returned in the responses to
create and list commands. You provide it to operations like update and
delete.}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.

Retrieves the specified IPSet.
}
\section{Request syntax}{
\preformatted{svc$get_ip_set(
  Name = "string",
  Scope = "CLOUDFRONT"|"REGIONAL",
  Id = "string"
)
}
}

\keyword{internal}
