% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_logging_configuration}
\alias{wafv2_get_logging_configuration}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_get_logging_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the web ACL for which you want to get
the LoggingConfiguration.}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.

Returns the LoggingConfiguration for the specified web ACL.
}
\section{Request syntax}{
\preformatted{svc$get_logging_configuration(
  ResourceArn = "string"
)
}
}

\keyword{internal}
