% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_rate_based_statement_managed_keys}
\alias{wafv2_get_rate_based_statement_managed_keys}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_get_rate_based_statement_managed_keys(Scope, WebACLName, WebACLId,
  RuleName)
}
\arguments{
\item{Scope}{[required] Specifies whether this is for an AWS CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB) or an API Gateway stage.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{WebACLName}{[required] The name of the Web ACL. You cannot change the name of a Web ACL after
you create it.}

\item{WebACLId}{[required] The unique identifier for the Web ACL. This ID is returned in the
responses to create and list commands. You provide it to operations like
update and delete.}

\item{RuleName}{[required] The name of the rate-based rule to get the keys for.}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
}
\details{
Retrieves the keys that are currently blocked by a rate-based rule. The
maximum number of managed keys that can be blocked for a single
rate-based rule is 10,000. If more than 10,000 addresses exceed the rate
limit, those with the highest rates are blocked.
}
\section{Request syntax}{
\preformatted{svc$get_rate_based_statement_managed_keys(
  Scope = "CLOUDFRONT"|"REGIONAL",
  WebACLName = "string",
  WebACLId = "string",
  RuleName = "string"
)
}
}

\keyword{internal}
