% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_sampled_requests}
\alias{wafv2_get_sampled_requests}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_get_sampled_requests(WebAclArn, RuleMetricName, Scope, TimeWindow,
  MaxItems)
}
\arguments{
\item{WebAclArn}{[required] The Amazon resource name (ARN) of the \code{WebACL} for which you want a
sample of requests.}

\item{RuleMetricName}{[required] The metric name assigned to the \code{Rule} or \code{RuleGroup} for which you want
a sample of requests.}

\item{Scope}{[required] Specifies whether this is for an AWS CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{TimeWindow}{[required] The start date and time and the end date and time of the range for which
you want \code{GetSampledRequests} to return a sample of requests. You must
specify the times in Coordinated Universal Time (UTC) format. UTC format
includes the special designator, \code{Z}. For example,
\code{"2016-09-27T14:50Z"}. You can specify any time range in the previous
three hours.}

\item{MaxItems}{[required] The number of requests that you want AWS WAF to return from among the
first 5,000 requests that your AWS resource received during the time
range. If your resource received fewer requests than the value of
\code{MaxItems}, \code{GetSampledRequests} returns information about all of them.}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.

Gets detailed information about a specified number of requests--a
sample--that AWS WAF randomly selects from among the first 5,000
requests that your AWS resource received during a time range that you
choose. You can specify a sample size of up to 500 requests, and you can
specify any time range in the previous three hours.

\code{GetSampledRequests} returns a time range, which is usually the time
range that you specified. However, if your resource (such as a
CloudFront distribution) received 5,000 requests before the specified
time range elapsed, \code{GetSampledRequests} returns an updated time range.
This new time range indicates the actual period during which AWS WAF
selected the requests in the sample.
}
\section{Request syntax}{
\preformatted{svc$get_sampled_requests(
  WebAclArn = "string",
  RuleMetricName = "string",
  Scope = "CLOUDFRONT"|"REGIONAL",
  TimeWindow = list(
    StartTime = as.POSIXct(
      "2015-01-01"
    ),
    EndTime = as.POSIXct(
      "2015-01-01"
    )
  ),
  MaxItems = 123
)
}
}

\keyword{internal}
