% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_web_acl_for_resource}
\alias{wafv2_get_web_acl_for_resource}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_get_web_acl_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN (Amazon Resource Name) of the resource.}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
}
\details{
Retrieves the WebACL for the specified resource.
}
\section{Request syntax}{
\preformatted{svc$get_web_acl_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
