% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_list_rule_groups}
\alias{wafv2_list_rule_groups}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_list_rule_groups(Scope, NextMarker, Limit)
}
\arguments{
\item{Scope}{[required] Specifies whether this is for an AWS CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{NextMarker}{When you request a list of objects with a \code{Limit} setting, if the number
of objects that are still available for retrieval exceeds the limit, AWS
WAF returns a \code{NextMarker} value in the response. To retrieve the next
batch of objects, provide the marker from the prior call in your next
request.}

\item{Limit}{The maximum number of objects that you want AWS WAF to return for this
request. If more objects are available, in the response, AWS WAF
provides a \code{NextMarker} value that you can use in a subsequent call to
get the next batch of objects.}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.

Retrieves an array of RuleGroupSummary objects for the rule groups that
you manage.
}
\section{Request syntax}{
\preformatted{svc$list_rule_groups(
  Scope = "CLOUDFRONT"|"REGIONAL",
  NextMarker = "string",
  Limit = 123
)
}
}

\keyword{internal}
