% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_workload}
\alias{wellarchitected_create_workload}
\title{Create a new workload}
\usage{
wellarchitected_create_workload(WorkloadName, Description, Environment,
  AccountIds, AwsRegions, NonAwsRegions, PillarPriorities,
  ArchitecturalDesign, ReviewOwner, IndustryType, Industry, Lenses, Notes,
  ClientRequestToken)
}
\arguments{
\item{WorkloadName}{[required]}

\item{Description}{[required]}

\item{Environment}{[required]}

\item{AccountIds}{}

\item{AwsRegions}{}

\item{NonAwsRegions}{}

\item{PillarPriorities}{}

\item{ArchitecturalDesign}{}

\item{ReviewOwner}{[required]}

\item{IndustryType}{}

\item{Industry}{}

\item{Lenses}{[required]}

\item{Notes}{}

\item{ClientRequestToken}{[required]}
}
\description{
Create a new workload.

The owner of a workload can share the workload with other AWS accounts
and IAM users in the same AWS Region. Only the owner of a workload can
delete it.

For more information, see \href{https://docs.aws.amazon.com/wellarchitected/latest/userguide/define-workload.html}{Defining a Workload}
in the \emph{AWS Well-Architected Tool User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_workload(
  WorkloadName = "string",
  Description = "string",
  Environment = "PRODUCTION"|"PREPRODUCTION",
  AccountIds = list(
    "string"
  ),
  AwsRegions = list(
    "string"
  ),
  NonAwsRegions = list(
    "string"
  ),
  PillarPriorities = list(
    "string"
  ),
  ArchitecturalDesign = "string",
  ReviewOwner = "string",
  IndustryType = "string",
  Industry = "string",
  Lenses = list(
    "string"
  ),
  Notes = "string",
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
