% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_workload_share}
\alias{wellarchitected_create_workload_share}
\title{Create a workload share}
\usage{
wellarchitected_create_workload_share(WorkloadId, SharedWith,
  PermissionType, ClientRequestToken)
}
\arguments{
\item{WorkloadId}{[required]}

\item{SharedWith}{[required]}

\item{PermissionType}{[required]}

\item{ClientRequestToken}{[required]}
}
\description{
Create a workload share.

The owner of a workload can share it with other AWS accounts and IAM
users in the same AWS Region. Shared access to a workload is not removed
until the workload invitation is deleted.

For more information, see \href{https://docs.aws.amazon.com/wellarchitected/latest/userguide/workloads-sharing.html}{Sharing a Workload}
in the \emph{AWS Well-Architected Tool User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_workload_share(
  WorkloadId = "string",
  SharedWith = "string",
  PermissionType = "READONLY"|"CONTRIBUTOR",
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
