% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_disassociate_lenses}
\alias{wellarchitected_disassociate_lenses}
\title{Disassociate a lens from a workload}
\usage{
wellarchitected_disassociate_lenses(WorkloadId, LensAliases)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAliases}{[required]}
}
\description{
Disassociate a lens from a workload.

The AWS Well-Architected Framework lens (\code{wellarchitected}) cannot be
removed from a workload.
}
\section{Request syntax}{
\preformatted{svc$disassociate_lenses(
  WorkloadId = "string",
  LensAliases = list(
    "string"
  )
)
}
}

\keyword{internal}
