% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_answer}
\alias{wellarchitected_get_answer}
\title{Get lens review}
\usage{
wellarchitected_get_answer(WorkloadId, LensAlias, QuestionId,
  MilestoneNumber)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{QuestionId}{[required]}

\item{MilestoneNumber}{}
}
\description{
Get lens review.
}
\section{Request syntax}{
\preformatted{svc$get_answer(
  WorkloadId = "string",
  LensAlias = "string",
  QuestionId = "string",
  MilestoneNumber = 123
)
}
}

\keyword{internal}
