% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_lens_review}
\alias{wellarchitected_get_lens_review}
\title{Get lens review}
\usage{
wellarchitected_get_lens_review(WorkloadId, LensAlias, MilestoneNumber)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{MilestoneNumber}{}
}
\description{
Get lens review.
}
\section{Request syntax}{
\preformatted{svc$get_lens_review(
  WorkloadId = "string",
  LensAlias = "string",
  MilestoneNumber = 123
)
}
}

\keyword{internal}
