% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_lens_version_difference}
\alias{wellarchitected_get_lens_version_difference}
\title{Get lens version differences}
\usage{
wellarchitected_get_lens_version_difference(LensAlias, BaseLensVersion)
}
\arguments{
\item{LensAlias}{[required]}

\item{BaseLensVersion}{[required] The base version of the lens.}
}
\description{
Get lens version differences.
}
\section{Request syntax}{
\preformatted{svc$get_lens_version_difference(
  LensAlias = "string",
  BaseLensVersion = "string"
)
}
}

\keyword{internal}
