% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_milestone}
\alias{wellarchitected_get_milestone}
\title{Get a milestone for an existing workload}
\usage{
wellarchitected_get_milestone(WorkloadId, MilestoneNumber)
}
\arguments{
\item{WorkloadId}{[required]}

\item{MilestoneNumber}{[required]}
}
\description{
Get a milestone for an existing workload.
}
\section{Request syntax}{
\preformatted{svc$get_milestone(
  WorkloadId = "string",
  MilestoneNumber = 123
)
}
}

\keyword{internal}
