% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_workload}
\alias{wellarchitected_get_workload}
\title{Get an existing workload}
\usage{
wellarchitected_get_workload(WorkloadId)
}
\arguments{
\item{WorkloadId}{[required]}
}
\description{
Get an existing workload.
}
\section{Request syntax}{
\preformatted{svc$get_workload(
  WorkloadId = "string"
)
}
}

\keyword{internal}
