% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_answers}
\alias{wellarchitected_list_answers}
\title{List of answers}
\usage{
wellarchitected_list_answers(WorkloadId, LensAlias, PillarId,
  MilestoneNumber, NextToken, MaxResults)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{PillarId}{}

\item{MilestoneNumber}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List of answers.
}
\section{Request syntax}{
\preformatted{svc$list_answers(
  WorkloadId = "string",
  LensAlias = "string",
  PillarId = "string",
  MilestoneNumber = 123,
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
