% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lens_review_improvements}
\alias{wellarchitected_list_lens_review_improvements}
\title{List lens review improvements}
\usage{
wellarchitected_list_lens_review_improvements(WorkloadId, LensAlias,
  PillarId, MilestoneNumber, NextToken, MaxResults)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{PillarId}{}

\item{MilestoneNumber}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List lens review improvements.
}
\section{Request syntax}{
\preformatted{svc$list_lens_review_improvements(
  WorkloadId = "string",
  LensAlias = "string",
  PillarId = "string",
  MilestoneNumber = 123,
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
