% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lens_reviews}
\alias{wellarchitected_list_lens_reviews}
\title{List lens reviews}
\usage{
wellarchitected_list_lens_reviews(WorkloadId, MilestoneNumber,
  NextToken, MaxResults)
}
\arguments{
\item{WorkloadId}{[required]}

\item{MilestoneNumber}{}

\item{NextToken}{}

\item{MaxResults}{}
}
\description{
List lens reviews.
}
\section{Request syntax}{
\preformatted{svc$list_lens_reviews(
  WorkloadId = "string",
  MilestoneNumber = 123,
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
