% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lenses}
\alias{wellarchitected_list_lenses}
\title{List the available lenses}
\usage{
wellarchitected_list_lenses(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{}

\item{MaxResults}{}
}
\description{
List the available lenses.
}
\section{Request syntax}{
\preformatted{svc$list_lenses(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
