% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_milestones}
\alias{wellarchitected_list_milestones}
\title{List all milestones for an existing workload}
\usage{
wellarchitected_list_milestones(WorkloadId, NextToken, MaxResults)
}
\arguments{
\item{WorkloadId}{[required]}

\item{NextToken}{}

\item{MaxResults}{}
}
\description{
List all milestones for an existing workload.
}
\section{Request syntax}{
\preformatted{svc$list_milestones(
  WorkloadId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
