% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_share_invitations}
\alias{wellarchitected_list_share_invitations}
\title{List the workload invitations}
\usage{
wellarchitected_list_share_invitations(WorkloadNamePrefix, NextToken,
  MaxResults)
}
\arguments{
\item{WorkloadNamePrefix}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List the workload invitations.
}
\section{Request syntax}{
\preformatted{svc$list_share_invitations(
  WorkloadNamePrefix = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
