% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_workloads}
\alias{wellarchitected_list_workloads}
\title{List workloads}
\usage{
wellarchitected_list_workloads(WorkloadNamePrefix, NextToken,
  MaxResults)
}
\arguments{
\item{WorkloadNamePrefix}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List workloads. Paginated.
}
\section{Request syntax}{
\preformatted{svc$list_workloads(
  WorkloadNamePrefix = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
