% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_answer}
\alias{wellarchitected_update_answer}
\title{Update the answer}
\usage{
wellarchitected_update_answer(WorkloadId, LensAlias, QuestionId,
  SelectedChoices, Notes, IsApplicable)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{QuestionId}{[required]}

\item{SelectedChoices}{}

\item{Notes}{}

\item{IsApplicable}{}
}
\description{
Update the answer.
}
\section{Request syntax}{
\preformatted{svc$update_answer(
  WorkloadId = "string",
  LensAlias = "string",
  QuestionId = "string",
  SelectedChoices = list(
    "string"
  ),
  Notes = "string",
  IsApplicable = TRUE|FALSE
)
}
}

\keyword{internal}
