% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_lens_review}
\alias{wellarchitected_update_lens_review}
\title{Update lens review}
\usage{
wellarchitected_update_lens_review(WorkloadId, LensAlias, LensNotes,
  PillarNotes)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{LensNotes}{}

\item{PillarNotes}{}
}
\description{
Update lens review.
}
\section{Request syntax}{
\preformatted{svc$update_lens_review(
  WorkloadId = "string",
  LensAlias = "string",
  LensNotes = "string",
  PillarNotes = list(
    "string"
  )
)
}
}

\keyword{internal}
