% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_workload}
\alias{wellarchitected_update_workload}
\title{Update an existing workload}
\usage{
wellarchitected_update_workload(WorkloadId, WorkloadName, Description,
  Environment, AccountIds, AwsRegions, NonAwsRegions, PillarPriorities,
  ArchitecturalDesign, ReviewOwner, IsReviewOwnerUpdateAcknowledged,
  IndustryType, Industry, Notes, ImprovementStatus)
}
\arguments{
\item{WorkloadId}{[required]}

\item{WorkloadName}{}

\item{Description}{}

\item{Environment}{}

\item{AccountIds}{}

\item{AwsRegions}{}

\item{NonAwsRegions}{}

\item{PillarPriorities}{}

\item{ArchitecturalDesign}{}

\item{ReviewOwner}{}

\item{IsReviewOwnerUpdateAcknowledged}{Flag indicating whether the workload owner has acknowledged that the
\emph{Review owner} field is required.

If a \strong{Review owner} is not added to the workload within 60 days of
acknowledgement, access to the workload is restricted until an owner is
added.}

\item{IndustryType}{}

\item{Industry}{}

\item{Notes}{}

\item{ImprovementStatus}{}
}
\description{
Update an existing workload.
}
\section{Request syntax}{
\preformatted{svc$update_workload(
  WorkloadId = "string",
  WorkloadName = "string",
  Description = "string",
  Environment = "PRODUCTION"|"PREPRODUCTION",
  AccountIds = list(
    "string"
  ),
  AwsRegions = list(
    "string"
  ),
  NonAwsRegions = list(
    "string"
  ),
  PillarPriorities = list(
    "string"
  ),
  ArchitecturalDesign = "string",
  ReviewOwner = "string",
  IsReviewOwnerUpdateAcknowledged = TRUE|FALSE,
  IndustryType = "string",
  Industry = "string",
  Notes = "string",
  ImprovementStatus = "NOT_APPLICABLE"|"NOT_STARTED"|"IN_PROGRESS"|"COMPLETE"|"RISK_ACKNOWLEDGED"
)
}
}

\keyword{internal}
