% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_workload_share}
\alias{wellarchitected_update_workload_share}
\title{Update a workload share}
\usage{
wellarchitected_update_workload_share(ShareId, WorkloadId,
  PermissionType)
}
\arguments{
\item{ShareId}{[required]}

\item{WorkloadId}{[required]}

\item{PermissionType}{[required]}
}
\description{
Update a workload share.
}
\section{Request syntax}{
\preformatted{svc$update_workload_share(
  ShareId = "string",
  WorkloadId = "string",
  PermissionType = "READONLY"|"CONTRIBUTOR"
)
}
}

\keyword{internal}
