% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_describe_comments}
\alias{workdocs_describe_comments}
\title{List all the comments for the specified document version}
\usage{
workdocs_describe_comments(AuthenticationToken, DocumentId, VersionId,
  Limit, Marker)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The ID of the document version.}

\item{Limit}{The maximum number of items to return.}

\item{Marker}{The marker for the next set of results. This marker was received from a
previous call.}
}
\description{
List all the comments for the specified document version.
}
\section{Request syntax}{
\preformatted{svc$describe_comments(
  AuthenticationToken = "string",
  DocumentId = "string",
  VersionId = "string",
  Limit = 123,
  Marker = "string"
)
}
}

\keyword{internal}
