% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_describe_document_versions}
\alias{workdocs_describe_document_versions}
\title{Retrieves the document versions for the specified document}
\usage{
workdocs_describe_document_versions(AuthenticationToken, DocumentId,
  Marker, Limit, Include, Fields)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using
administrative API actions, as in accessing the API using AWS
credentials.}

\item{DocumentId}{[required] The ID of the document.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{Limit}{The maximum number of versions to return with this call.}

\item{Include}{A comma-separated list of values. Specify "INITIALIZED" to include
incomplete versions.}

\item{Fields}{Specify "SOURCE" to include initialized versions and a URL for the
source document.}
}
\description{
Retrieves the document versions for the specified document.
}
\details{
By default, only active versions are returned.
}
\section{Request syntax}{
\preformatted{svc$describe_document_versions(
  AuthenticationToken = "string",
  DocumentId = "string",
  Marker = "string",
  Limit = 123,
  Include = "string",
  Fields = "string"
)
}
}

\keyword{internal}
