% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_remove_all_resource_permissions}
\alias{workdocs_remove_all_resource_permissions}
\title{Removes all the permissions from the specified resource}
\usage{
workdocs_remove_all_resource_permissions(AuthenticationToken,
  ResourceId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using
administrative API actions, as in accessing the API using AWS
credentials.}

\item{ResourceId}{[required] The ID of the resource.}
}
\description{
Removes all the permissions from the specified resource.
}
\section{Request syntax}{
\preformatted{svc$remove_all_resource_permissions(
  AuthenticationToken = "string",
  ResourceId = "string"
)
}
}

\keyword{internal}
