% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_describe_audit_stream_configuration}
\alias{worklink_describe_audit_stream_configuration}
\title{Describes the configuration for delivering audit streams to the customer
account}
\usage{
worklink_describe_audit_stream_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\description{
Describes the configuration for delivering audit streams to the customer
account.
}
\section{Request syntax}{
\preformatted{worklink$describe_audit_stream_configuration(
  FleetArn = "string"
)
}
}

\keyword{internal}
