% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_update_device_policy_configuration}
\alias{worklink_update_device_policy_configuration}
\title{Updates the device policy configuration for the fleet}
\usage{
worklink_update_device_policy_configuration(FleetArn,
  DeviceCaCertificate)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{DeviceCaCertificate}{The certificate chain, including intermediate certificates and the root
certificate authority certificate used to issue device certificates.}
}
\description{
Updates the device policy configuration for the fleet.
}
\section{Request syntax}{
\preformatted{svc$update_device_policy_configuration(
  FleetArn = "string",
  DeviceCaCertificate = "string"
)
}
}

\keyword{internal}
