% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_mailbox_permissions}
\alias{workmail_delete_mailbox_permissions}
\title{Deletes permissions granted to a user or group}
\usage{
workmail_delete_mailbox_permissions(OrganizationId, EntityId, GranteeId)
}
\arguments{
\item{OrganizationId}{[required] The identifier of the organization under which the entity (user or
group) exists.}

\item{EntityId}{[required] The identifier of the entity (user or group) for which to delete mailbox
permissions.}

\item{GranteeId}{[required] The identifier of the entity (user or group) for which to delete granted
permissions.}
}
\description{
Deletes permissions granted to a user or group.
}
\section{Request syntax}{
\preformatted{workmail$delete_mailbox_permissions(
  OrganizationId = "string",
  EntityId = "string",
  GranteeId = "string"
)
}
}

\keyword{internal}
