% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_user}
\alias{workmail_delete_user}
\title{Deletes a user from Amazon WorkMail and all subsequent systems}
\usage{
workmail_delete_user(OrganizationId, UserId)
}
\arguments{
\item{OrganizationId}{[required] The organization that contains the user.}

\item{UserId}{[required] The identifier of the user to be deleted.}
}
\description{
Deletes a user from Amazon WorkMail and all subsequent systems. The
action can't be undone. The mailbox is kept as-is for a minimum of 30
days, without any means to restore it.
}
\section{Request syntax}{
\preformatted{workmail$delete_user(
  OrganizationId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
