% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_user}
\alias{workmail_describe_user}
\title{Provides information regarding the user}
\usage{
workmail_describe_user(OrganizationId, UserId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the user exists.}

\item{UserId}{[required] The identifier for the user to be described.}
}
\description{
Provides information regarding the user.
}
\section{Request syntax}{
\preformatted{workmail$describe_user(
  OrganizationId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
