% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_mailbox_export_jobs}
\alias{workmail_list_mailbox_export_jobs}
\title{Lists the mailbox export jobs started for the specified organization
within the last seven days}
\usage{
workmail_list_mailbox_export_jobs(OrganizationId, NextToken, MaxResults)
}
\arguments{
\item{OrganizationId}{[required] The organization ID.}

\item{NextToken}{The token to use to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Lists the mailbox export jobs started for the specified organization
within the last seven days.
}
\section{Request syntax}{
\preformatted{svc$list_mailbox_export_jobs(
  OrganizationId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
