% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_put_mailbox_permissions}
\alias{workmail_put_mailbox_permissions}
\title{Sets permissions for a user or group}
\usage{
workmail_put_mailbox_permissions(OrganizationId, EntityId, GranteeId,
  PermissionValues)
}
\arguments{
\item{OrganizationId}{[required] The identifier of the organization under which the entity (user or
group) exists.}

\item{EntityId}{[required] The identifier of the entity (user or group) for which to update mailbox
permissions.}

\item{GranteeId}{[required] The identifier of the entity (user or group) to which to grant the
permissions.}

\item{PermissionValues}{[required] The permissions granted to the grantee. SEND\_AS allows the grantee to
send email as the owner of the mailbox (the grantee is not mentioned on
these emails). SEND\_ON\_BEHALF allows the grantee to send email on
behalf of the owner of the mailbox (the grantee is not mentioned as the
physical sender of these emails). FULL\_ACCESS allows the grantee full
access to the mailbox, irrespective of other folder-level permissions
set on the mailbox.}
}
\description{
Sets permissions for a user or group. This replaces any pre-existing
permissions set for the entity.
}
\section{Request syntax}{
\preformatted{workmail$put_mailbox_permissions(
  OrganizationId = "string",
  EntityId = "string",
  GranteeId = "string",
  PermissionValues = list(
    "FULL_ACCESS"|"SEND_AS"|"SEND_ON_BEHALF"
  )
)
}
}

\keyword{internal}
