% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_delete_tags}
\alias{workspaces_delete_tags}
\title{Deletes the specified tags from the specified WorkSpace}
\usage{
workspaces_delete_tags(ResourceId, TagKeys)
}
\arguments{
\item{ResourceId}{[required] The identifier of the WorkSpace. To find this ID, use
DescribeWorkspaces.}

\item{TagKeys}{[required] The tag keys.}
}
\description{
Deletes the specified tags from the specified WorkSpace.
}
\section{Request syntax}{
\preformatted{workspaces$delete_tags(
  ResourceId = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
