% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_account}
\alias{workspaces_describe_account}
\title{Retrieves a list that describes the configuration of Bring Your Own
License (BYOL) for the specified account}
\usage{
workspaces_describe_account()
}
\description{
Retrieves a list that describes the configuration of Bring Your Own
License (BYOL) for the specified account.
}
\section{Request syntax}{
\preformatted{svc$describe_account()
}
}

\keyword{internal}
