% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_client_properties}
\alias{workspaces_describe_client_properties}
\title{Retrieves a list that describes one or more specified Amazon WorkSpaces
clients}
\usage{
workspaces_describe_client_properties(ResourceIds)
}
\arguments{
\item{ResourceIds}{[required] The resource identifiers, in the form of directory IDs.}
}
\description{
Retrieves a list that describes one or more specified Amazon WorkSpaces
clients.
}
\section{Request syntax}{
\preformatted{workspaces$describe_client_properties(
  ResourceIds = list(
    "string"
  )
)
}
}

\keyword{internal}
