% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_modify_workspace_creation_properties}
\alias{workspaces_modify_workspace_creation_properties}
\title{Modify the default properties used to create WorkSpaces}
\usage{
workspaces_modify_workspace_creation_properties(ResourceId,
  WorkspaceCreationProperties)
}
\arguments{
\item{ResourceId}{[required] The identifier of the directory.}

\item{WorkspaceCreationProperties}{[required] The default properties for creating WorkSpaces.}
}
\description{
Modify the default properties used to create WorkSpaces.
}
\section{Request syntax}{
\preformatted{svc$modify_workspace_creation_properties(
  ResourceId = "string",
  WorkspaceCreationProperties = list(
    EnableInternetAccess = TRUE|FALSE,
    DefaultOu = "string",
    CustomSecurityGroupId = "string",
    UserEnabledAsLocalAdministrator = TRUE|FALSE,
    EnableMaintenanceMode = TRUE|FALSE
  )
)
}
}

\keyword{internal}
