% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_reboot_workspaces}
\alias{workspaces_reboot_workspaces}
\title{Reboots the specified WorkSpaces}
\usage{
workspaces_reboot_workspaces(RebootWorkspaceRequests)
}
\arguments{
\item{RebootWorkspaceRequests}{[required] The WorkSpaces to reboot. You can specify up to 25 WorkSpaces.}
}
\description{
Reboots the specified WorkSpaces.
}
\details{
You cannot reboot a WorkSpace unless its state is \code{AVAILABLE} or
\code{UNHEALTHY}.

This operation is asynchronous and returns before the WorkSpaces have
rebooted.
}
\section{Request syntax}{
\preformatted{svc$reboot_workspaces(
  RebootWorkspaceRequests = list(
    list(
      WorkspaceId = "string"
    )
  )
)
}
}

\keyword{internal}
