% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmpca_operations.R
\name{acmpca_create_permission}
\alias{acmpca_create_permission}
\title{Grants one or more permissions on a private CA to the AWS Certificate
Manager (ACM) service principal (acm}
\usage{
acmpca_create_permission(CertificateAuthorityArn, Principal,
  SourceAccount, Actions)
}
\arguments{
\item{CertificateAuthorityArn}{[required] The Amazon Resource Name (ARN) of the CA that grants the permissions.
You can find the ARN by calling the
\href{https://docs.aws.amazon.com/acm-pca/latest/APIReference/API_ListCertificateAuthorities.html}{ListCertificateAuthorities}
action. This must have the following form:

\verb{arn:aws:acm-pca:<i>region</i>:<i>account</i>:certificate-authority/<i>12345678-1234-1234-1234-123456789012</i> }.}

\item{Principal}{[required] The AWS service or identity that receives the permission. At this time,
the only valid principal is \code{acm.amazonaws.com}.}

\item{SourceAccount}{The ID of the calling account.}

\item{Actions}{[required] The actions that the specified AWS service principal can use. These
include \code{IssueCertificate}, \code{GetCertificate}, and \code{ListPermissions}.}
}
\description{
Grants one or more permissions on a private CA to the AWS Certificate
Manager (ACM) service principal (\code{acm.amazonaws.com}). These permissions
allow ACM to issue and renew ACM certificates that reside in the same
AWS account as the CA.
}
\details{
You can list current permissions with the
\href{https://docs.aws.amazon.com/acm-pca/latest/APIReference/API_ListPermissions.html}{ListPermissions}
action and revoke them with the
\href{https://docs.aws.amazon.com/acm-pca/latest/APIReference/API_DeletePermission.html}{DeletePermission}
action.

\strong{About Permissions}
\itemize{
\item If the private CA and the certificates it issues reside in the same
account, you can use \code{CreatePermission} to grant permissions for ACM
to carry out automatic certificate renewals.
\item For automatic certificate renewal to succeed, the ACM service
principal needs permissions to create, retrieve, and list
certificates.
\item If the private CA and the ACM certificates reside in different
accounts, then permissions cannot be used to enable automatic
renewals. Instead, the ACM certificate owner must set up a
resource-based policy to enable cross-account issuance and renewals.
For more information, see \href{https://docs.aws.amazon.com/acm-pca/latest/userguide/pca-rbp.html}{Using a Resource Based Policy with ACM Private CA}.
}
}
\section{Request syntax}{
\preformatted{svc$create_permission(
  CertificateAuthorityArn = "string",
  Principal = "string",
  SourceAccount = "string",
  Actions = list(
    "IssueCertificate"|"GetCertificate"|"ListPermissions"
  )
)
}
}

\keyword{internal}
