% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{amplify}
\alias{amplify}
\title{AWS Amplify}
\usage{
amplify()
}
\description{
Amplify is a fully managed continuous deployment and hosting service for
modern web apps.
}
\section{Operations}{

\tabular{ll}{
\link[=amplify_create_app]{create_app} \tab Creates a new Amplify App \cr
\link[=amplify_create_branch]{create_branch} \tab Creates a new Branch for an Amplify App \cr
\link[=amplify_create_domain_association]{create_domain_association} \tab Create a new DomainAssociation on an App \cr
\link[=amplify_delete_app]{delete_app} \tab Delete an existing Amplify App by appId \cr
\link[=amplify_delete_branch]{delete_branch} \tab Deletes a branch for an Amplify App \cr
\link[=amplify_delete_domain_association]{delete_domain_association} \tab Deletes a DomainAssociation \cr
\link[=amplify_delete_job]{delete_job} \tab Delete a job, for an Amplify branch, part of Amplify App \cr
\link[=amplify_get_app]{get_app} \tab Retrieves an existing Amplify App by appId \cr
\link[=amplify_get_branch]{get_branch} \tab Retrieves a branch for an Amplify App \cr
\link[=amplify_get_domain_association]{get_domain_association} \tab Retrieves domain info that corresponds to an appId and domainName \cr
\link[=amplify_get_job]{get_job} \tab Get a job for a branch, part of an Amplify App \cr
\link[=amplify_list_apps]{list_apps} \tab Lists existing Amplify Apps \cr
\link[=amplify_list_branches]{list_branches} \tab Lists branches for an Amplify App \cr
\link[=amplify_list_domain_associations]{list_domain_associations} \tab List domains with an app \cr
\link[=amplify_list_jobs]{list_jobs} \tab List Jobs for a branch, part of an Amplify App \cr
\link[=amplify_start_job]{start_job} \tab Starts a new job for a branch, part of an Amplify App \cr
\link[=amplify_stop_job]{stop_job} \tab Stop a job that is in progress, for an Amplify branch, part of Amplify App\cr
\link[=amplify_update_app]{update_app} \tab Updates an existing Amplify App \cr
\link[=amplify_update_branch]{update_branch} \tab Updates a branch for an Amplify App \cr
\link[=amplify_update_domain_association]{update_domain_association} \tab Create a new DomainAssociation on an App
}
}

\examples{
\donttest{svc <- amplify()
svc$create_app(
  Foo = 123
)}

}
