% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_create_app}
\alias{amplify_create_app}
\title{Creates a new Amplify App}
\usage{
amplify_create_app(name, description, repository, platform,
  iamServiceRoleArn, oauthToken, environmentVariables,
  enableBranchAutoBuild, enableBasicAuth, basicAuthCredentials,
  customRules, tags, buildSpec)
}
\arguments{
\item{name}{[required] Name for the Amplify App}

\item{description}{Description for an Amplify App}

\item{repository}{[required] Repository for an Amplify App}

\item{platform}{[required] Platform / framework for an Amplify App}

\item{iamServiceRoleArn}{AWS IAM service role for an Amplify App}

\item{oauthToken}{[required] OAuth token for 3rd party source control system for an Amplify App, used
to create webhook and read-only deploy key. OAuth token is not stored.}

\item{environmentVariables}{Environment variables map for an Amplify App.}

\item{enableBranchAutoBuild}{Enable the auto building of branches for an Amplify App.}

\item{enableBasicAuth}{Enable Basic Authorization for an Amplify App, this will apply to all
branches part of this App.}

\item{basicAuthCredentials}{Credentials for Basic Authorization for an Amplify App.}

\item{customRules}{Custom rewrite / redirect rules for an Amplify App.}

\item{tags}{Tag for an Amplify App}

\item{buildSpec}{BuildSpec for an Amplify App}
}
\description{
Creates a new Amplify App.
}
\section{Request syntax}{
\preformatted{amplify$create_app(
  name = "string",
  description = "string",
  repository = "string",
  platform = "IOS"|"ANDROID"|"WEB"|"REACT_NATIVE",
  iamServiceRoleArn = "string",
  oauthToken = "string",
  environmentVariables = list(
    "string"
  ),
  enableBranchAutoBuild = TRUE|FALSE,
  enableBasicAuth = TRUE|FALSE,
  basicAuthCredentials = "string",
  customRules = list(
    list(
      source = "string",
      target = "string",
      status = "string",
      condition = "string"
    )
  ),
  tags = list(
    "string"
  ),
  buildSpec = "string"
)
}
}

\keyword{internal}
