% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_generate_access_logs}
\alias{amplify_generate_access_logs}
\title{Returns the website access logs for a specific time range using a
presigned URL}
\usage{
amplify_generate_access_logs(startTime, endTime, domainName, appId)
}
\arguments{
\item{startTime}{The time at which the logs should start. The time range specified is
inclusive of the start time.}

\item{endTime}{The time at which the logs should end. The time range specified is
inclusive of the end time.}

\item{domainName}{[required] The name of the domain.}

\item{appId}{[required] The unique ID for an Amplify app.}
}
\description{
Returns the website access logs for a specific time range using a
presigned URL.
}
\section{Request syntax}{
\preformatted{svc$generate_access_logs(
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  domainName = "string",
  appId = "string"
)
}
}

\keyword{internal}
