% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_webhook}
\alias{amplify_get_webhook}
\title{Returns the webhook information that corresponds to a specified webhook
ID}
\usage{
amplify_get_webhook(webhookId)
}
\arguments{
\item{webhookId}{[required] The unique ID for a webhook.}
}
\description{
Returns the webhook information that corresponds to a specified webhook
ID.
}
\section{Request syntax}{
\preformatted{svc$get_webhook(
  webhookId = "string"
)
}
}

\keyword{internal}
