% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_update_branch}
\alias{amplify_update_branch}
\title{Updates a branch for an Amplify App}
\usage{
amplify_update_branch(appId, branchName, description, framework, stage,
  enableNotification, enableAutoBuild, environmentVariables,
  basicAuthCredentials, enableBasicAuth, buildSpec, ttl, displayName,
  enablePullRequestPreview, pullRequestEnvironmentName,
  backendEnvironmentArn)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch.}

\item{description}{Description for the branch.}

\item{framework}{Framework for the branch.}

\item{stage}{Stage for the branch.}

\item{enableNotification}{Enables notifications for the branch.}

\item{enableAutoBuild}{Enables auto building for the branch.}

\item{environmentVariables}{Environment Variables for the branch.}

\item{basicAuthCredentials}{Basic Authorization credentials for the branch.}

\item{enableBasicAuth}{Enables Basic Auth for the branch.}

\item{buildSpec}{BuildSpec for the branch.}

\item{ttl}{The content TTL for the website in seconds.}

\item{displayName}{Display name for a branch, will use as the default domain prefix.}

\item{enablePullRequestPreview}{Enables Pull Request Preview for this branch.}

\item{pullRequestEnvironmentName}{The Amplify Environment name for the pull request.}

\item{backendEnvironmentArn}{ARN for a Backend Environment, part of an Amplify App.}
}
\description{
Updates a branch for an Amplify App.
}
\section{Request syntax}{
\preformatted{svc$update_branch(
  appId = "string",
  branchName = "string",
  description = "string",
  framework = "string",
  stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL"|"PULL_REQUEST",
  enableNotification = TRUE|FALSE,
  enableAutoBuild = TRUE|FALSE,
  environmentVariables = list(
    "string"
  ),
  basicAuthCredentials = "string",
  enableBasicAuth = TRUE|FALSE,
  buildSpec = "string",
  ttl = "string",
  displayName = "string",
  enablePullRequestPreview = TRUE|FALSE,
  pullRequestEnvironmentName = "string",
  backendEnvironmentArn = "string"
)
}
}

\keyword{internal}
