% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewaymanagementapi_service.R
\name{apigatewaymanagementapi}
\alias{apigatewaymanagementapi}
\title{AmazonApiGatewayManagementApi}
\usage{
apigatewaymanagementapi()
}
\description{
The Amazon API Gateway Management API allows you to directly manage
runtime aspects of your deployed APIs. To use it, you must explicitly
set the SDK's endpoint to point to the endpoint of your deployed API.
The endpoint will be of the form
https://{api-id}.execute-api.{region}.amazonaws.com/{stage}, or will be
the endpoint corresponding to your API's custom domain and base path,
if applicable.
}
\section{Operations}{

\tabular{ll}{
\link[=apigatewaymanagementapi_post_to_connection]{post_to_connection} \tab Sends the provided data to the specified connection
}
}

\examples{
\donttest{svc <- apigatewaymanagementapi()
svc$post_to_connection(
  Foo = 123
)}

}
