% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_domain_name}
\alias{apigatewayv2_create_domain_name}
\title{Creates a domain name}
\usage{
apigatewayv2_create_domain_name(DomainName, DomainNameConfigurations)
}
\arguments{
\item{DomainName}{[required] The domain name.}

\item{DomainNameConfigurations}{The domain name configurations.}
}
\description{
Creates a domain name.
}
\section{Request syntax}{
\preformatted{svc$create_domain_name(
  DomainName = "string",
  DomainNameConfigurations = list(
    list(
      ApiGatewayDomainName = "string",
      CertificateArn = "string",
      CertificateName = "string",
      CertificateUploadDate = as.POSIXct(
        "2015-01-01"
      ),
      EndpointType = "REGIONAL"|"EDGE",
      HostedZoneId = "string"
    )
  )
)
}
}

\keyword{internal}
