% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appconfig_operations.R
\name{appconfig_create_deployment_strategy}
\alias{appconfig_create_deployment_strategy}
\title{A deployment strategy defines important criteria for rolling out your
configuration to the designated targets}
\usage{
appconfig_create_deployment_strategy(Name, Description,
  DeploymentDurationInMinutes, FinalBakeTimeInMinutes, GrowthFactor,
  GrowthType, ReplicateTo, Tags)
}
\arguments{
\item{Name}{[required] A name for the deployment strategy.}

\item{Description}{A description of the deployment strategy.}

\item{DeploymentDurationInMinutes}{[required] Total amount of time for a deployment to last.}

\item{FinalBakeTimeInMinutes}{The amount of time AppConfig monitors for alarms before considering the
deployment to be complete and no longer eligible for automatic roll
back.}

\item{GrowthFactor}{[required] The percentage of targets to receive a deployed configuration during
each interval.}

\item{GrowthType}{The algorithm used to define how percentage grows over time.}

\item{ReplicateTo}{[required] Save the deployment strategy to a Systems Manager (SSM) document.}

\item{Tags}{Metadata to assign to the deployment strategy. Tags help organize and
categorize your AppConfig resources. Each tag consists of a key and an
optional value, both of which you define.}
}
\description{
A deployment strategy defines important criteria for rolling out your
configuration to the designated targets. A deployment strategy includes:
the overall duration required, a percentage of targets to receive the
deployment during each interval, an algorithm that defines how
percentage grows, and bake time.
}
\section{Request syntax}{
\preformatted{svc$create_deployment_strategy(
  Name = "string",
  Description = "string",
  DeploymentDurationInMinutes = 123,
  FinalBakeTimeInMinutes = 123,
  GrowthFactor = 123.0,
  GrowthType = "LINEAR",
  ReplicateTo = "NONE"|"SSM_DOCUMENT",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
