% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_delete_route}
\alias{appmesh_delete_route}
\title{Deletes an existing route}
\usage{
appmesh_delete_route(meshName, routeName, virtualRouterName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to delete the route in.}

\item{routeName}{[required] The name of the route to delete.}

\item{virtualRouterName}{[required] The name of the virtual router to delete the route in.}
}
\description{
Deletes an existing route.
}
\section{Request syntax}{
\preformatted{svc$delete_route(
  meshName = "string",
  routeName = "string",
  virtualRouterName = "string"
)
}
}

\keyword{internal}
