% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_service.R
\name{athena}
\alias{athena}
\title{Amazon Athena}
\usage{
athena()
}
\description{
Amazon Athena is an interactive query service that lets you use standard
SQL to analyze data directly in Amazon S3. You can point Athena at your
data in Amazon S3 and run ad-hoc queries and get results in seconds.
Athena is serverless, so there is no infrastructure to set up or manage.
You pay only for the queries you run. Athena scales
automatically---executing queries in parallel---so results are fast,
even with large datasets and complex queries. For more information, see
\href{http://docs.aws.amazon.com/athena/latest/ug/what-is.html}{What is Amazon Athena} in the
\emph{Amazon Athena User Guide}.

If you connect to Athena using the JDBC driver, use version 1.1.0 of the
driver or later with the Amazon Athena API. Earlier version drivers do
not support the API. For more information and to download the driver,
see \href{https://docs.aws.amazon.com/athena/latest/ug/connect-with-jdbc.html}{Accessing Amazon Athena with JDBC}.

For code samples using the AWS SDK for Java, see \href{https://docs.aws.amazon.com/athena/latest/ug/code-samples.html}{Examples and Code Samples}
in the \emph{Amazon Athena User Guide}.
}
\section{Example}{
\preformatted{athena <- paws::athena()
athena$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=athena_batch_get_named_query]{batch_get_named_query} \tab Returns the details of a single named query or a list of up to 50 queries, which you provide as an array of query ID strings \cr
\link[=athena_batch_get_query_execution]{batch_get_query_execution} \tab Returns the details of a single query execution or a list of up to 50 query executions, which you provide as an array of query execution ID strings\cr
\link[=athena_create_named_query]{create_named_query} \tab Creates a named query \cr
\link[=athena_delete_named_query]{delete_named_query} \tab Deletes a named query \cr
\link[=athena_get_named_query]{get_named_query} \tab Returns information about a single query \cr
\link[=athena_get_query_execution]{get_query_execution} \tab Returns information about a single execution of a query \cr
\link[=athena_get_query_results]{get_query_results} \tab Returns the results of a single query execution specified by QueryExecutionId \cr
\link[=athena_list_named_queries]{list_named_queries} \tab Provides a list of all available query IDs \cr
\link[=athena_list_query_executions]{list_query_executions} \tab Provides a list of all available query execution IDs \cr
\link[=athena_start_query_execution]{start_query_execution} \tab Runs (executes) the SQL query statements contained in the Query string \cr
\link[=athena_stop_query_execution]{stop_query_execution} \tab Stops a query execution
}
}

