% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_query_executions}
\alias{athena_list_query_executions}
\title{Provides a list of all available query execution IDs}
\usage{
athena_list_query_executions(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token that specifies where to start pagination if a previous request
was truncated.}

\item{MaxResults}{The maximum number of query executions to return in this request.}
}
\description{
Provides a list of all available query execution IDs.
}
\details{
For code samples using the AWS SDK for Java, see \href{http://docs.aws.amazon.com/athena/latest/ug/code-samples.html}{Examples and Code Samples}
in the \emph{Amazon Athena User Guide}.
}
\section{Request syntax}{
\preformatted{athena$list_query_executions(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
