% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_service.R
\name{auditmanager}
\alias{auditmanager}
\title{AWS Audit Manager}
\usage{
auditmanager(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Welcome to the AWS Audit Manager API reference. This guide is for
developers who need detailed information about the AWS Audit Manager API
operations, data types, and errors.

AWS Audit Manager is a service that provides automated evidence
collection so that you can continuously audit your AWS usage, and assess
the effectiveness of your controls to better manage risk and simplify
compliance.

AWS Audit Manager provides pre-built frameworks that structure and
automate assessments for a given compliance standard. Frameworks include
a pre-built collection of controls with descriptions and testing
procedures, which are grouped according to the requirements of the
specified compliance standard or regulation. You can also customize
frameworks and controls to support internal audits with unique
requirements.

Use the following links to get started with the AWS Audit Manager API:
\itemize{
\item \href{https://docs.aws.amazon.com/audit-manager/latest/APIReference/API_Operations.html}{Actions}:
An alphabetical list of all AWS Audit Manager API operations.
\item \href{https://docs.aws.amazon.com/audit-manager/latest/APIReference/API_Types.html}{Data types}:
An alphabetical list of all AWS Audit Manager data types.
\item \href{https://docs.aws.amazon.com/audit-manager/latest/APIReference/CommonParameters.html}{Common parameters}:
Parameters that all Query operations can use.
\item \href{https://docs.aws.amazon.com/audit-manager/latest/APIReference/CommonErrors.html}{Common errors}:
Client and server errors that all operations can return.
}

If you're new to AWS Audit Manager, we recommend that you review the
\href{https://docs.aws.amazon.com/audit-manager/latest/userguide/what-is.html}{AWS Audit Manager User Guide}.
}
\section{Service syntax}{
\preformatted{svc <- auditmanager(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=auditmanager_associate_assessment_report_evidence_folder]{associate_assessment_report_evidence_folder} \tab Associates an evidence folder to the specified assessment report in AWS Audit Manager\cr
\link[=auditmanager_batch_associate_assessment_report_evidence]{batch_associate_assessment_report_evidence} \tab Associates a list of evidence to an assessment report in an AWS Audit Manager assessment\cr
\link[=auditmanager_batch_create_delegation_by_assessment]{batch_create_delegation_by_assessment} \tab Create a batch of delegations for a specified assessment in AWS Audit Manager\cr
\link[=auditmanager_batch_delete_delegation_by_assessment]{batch_delete_delegation_by_assessment} \tab Deletes the delegations in the specified AWS Audit Manager assessment\cr
\link[=auditmanager_batch_disassociate_assessment_report_evidence]{batch_disassociate_assessment_report_evidence} \tab Disassociates a list of evidence from the specified assessment report in AWS Audit Manager\cr
\link[=auditmanager_batch_import_evidence_to_assessment_control]{batch_import_evidence_to_assessment_control} \tab Uploads one or more pieces of evidence to the specified control in the assessment in AWS Audit Manager\cr
\link[=auditmanager_create_assessment]{create_assessment} \tab Creates an assessment in AWS Audit Manager\cr
\link[=auditmanager_create_assessment_framework]{create_assessment_framework} \tab Creates a custom framework in AWS Audit Manager\cr
\link[=auditmanager_create_assessment_report]{create_assessment_report} \tab Creates an assessment report for the specified assessment\cr
\link[=auditmanager_create_control]{create_control} \tab Creates a new custom control in AWS Audit Manager\cr
\link[=auditmanager_delete_assessment]{delete_assessment} \tab Deletes an assessment in AWS Audit Manager\cr
\link[=auditmanager_delete_assessment_framework]{delete_assessment_framework} \tab Deletes a custom framework in AWS Audit Manager\cr
\link[=auditmanager_delete_assessment_report]{delete_assessment_report} \tab Deletes an assessment report from an assessment in AWS Audit Manager\cr
\link[=auditmanager_delete_control]{delete_control} \tab Deletes a custom control in AWS Audit Manager\cr
\link[=auditmanager_deregister_account]{deregister_account} \tab Deregisters an account in AWS Audit Manager\cr
\link[=auditmanager_deregister_organization_admin_account]{deregister_organization_admin_account} \tab Deregisters the delegated AWS administrator account from the AWS organization\cr
\link[=auditmanager_disassociate_assessment_report_evidence_folder]{disassociate_assessment_report_evidence_folder} \tab Disassociates an evidence folder from the specified assessment report in AWS Audit Manager\cr
\link[=auditmanager_get_account_status]{get_account_status} \tab Returns the registration status of an account in AWS Audit Manager\cr
\link[=auditmanager_get_assessment]{get_assessment} \tab Returns an assessment from AWS Audit Manager\cr
\link[=auditmanager_get_assessment_framework]{get_assessment_framework} \tab Returns a framework from AWS Audit Manager\cr
\link[=auditmanager_get_assessment_report_url]{get_assessment_report_url} \tab Returns the URL of a specified assessment report in AWS Audit Manager\cr
\link[=auditmanager_get_change_logs]{get_change_logs} \tab Returns a list of changelogs from AWS Audit Manager\cr
\link[=auditmanager_get_control]{get_control} \tab Returns a control from AWS Audit Manager\cr
\link[=auditmanager_get_delegations]{get_delegations} \tab Returns a list of delegations from an audit owner to a delegate\cr
\link[=auditmanager_get_evidence]{get_evidence} \tab Returns evidence from AWS Audit Manager\cr
\link[=auditmanager_get_evidence_by_evidence_folder]{get_evidence_by_evidence_folder} \tab Returns all evidence from a specified evidence folder in AWS Audit Manager\cr
\link[=auditmanager_get_evidence_folder]{get_evidence_folder} \tab Returns an evidence folder from the specified assessment in AWS Audit Manager\cr
\link[=auditmanager_get_evidence_folders_by_assessment]{get_evidence_folders_by_assessment} \tab Returns the evidence folders from a specified assessment in AWS Audit Manager\cr
\link[=auditmanager_get_evidence_folders_by_assessment_control]{get_evidence_folders_by_assessment_control} \tab Returns a list of evidence folders associated with a specified control of an assessment in AWS Audit Manager\cr
\link[=auditmanager_get_organization_admin_account]{get_organization_admin_account} \tab Returns the name of the delegated AWS administrator account for the AWS organization\cr
\link[=auditmanager_get_services_in_scope]{get_services_in_scope} \tab Returns a list of the in-scope AWS services for the specified assessment\cr
\link[=auditmanager_get_settings]{get_settings} \tab Returns the settings for the specified AWS account\cr
\link[=auditmanager_list_assessment_frameworks]{list_assessment_frameworks} \tab Returns a list of the frameworks available in the AWS Audit Manager framework library\cr
\link[=auditmanager_list_assessment_reports]{list_assessment_reports} \tab Returns a list of assessment reports created in AWS Audit Manager\cr
\link[=auditmanager_list_assessments]{list_assessments} \tab Returns a list of current and past assessments from AWS Audit Manager\cr
\link[=auditmanager_list_controls]{list_controls} \tab Returns a list of controls from AWS Audit Manager\cr
\link[=auditmanager_list_keywords_for_data_source]{list_keywords_for_data_source} \tab Returns a list of keywords that pre-mapped to the specified control data source\cr
\link[=auditmanager_list_notifications]{list_notifications} \tab Returns a list of all AWS Audit Manager notifications\cr
\link[=auditmanager_list_tags_for_resource]{list_tags_for_resource} \tab Returns a list of tags for the specified resource in AWS Audit Manager\cr
\link[=auditmanager_register_account]{register_account} \tab Enables AWS Audit Manager for the specified AWS account\cr
\link[=auditmanager_register_organization_admin_account]{register_organization_admin_account} \tab Enables an AWS account within the organization as the delegated administrator for AWS Audit Manager\cr
\link[=auditmanager_tag_resource]{tag_resource} \tab Tags the specified resource in AWS Audit Manager\cr
\link[=auditmanager_untag_resource]{untag_resource} \tab Removes a tag from a resource in AWS Audit Manager\cr
\link[=auditmanager_update_assessment]{update_assessment} \tab Edits an AWS Audit Manager assessment\cr
\link[=auditmanager_update_assessment_control]{update_assessment_control} \tab Updates a control within an assessment in AWS Audit Manager\cr
\link[=auditmanager_update_assessment_control_set_status]{update_assessment_control_set_status} \tab Updates the status of a control set in an AWS Audit Manager assessment\cr
\link[=auditmanager_update_assessment_framework]{update_assessment_framework} \tab Updates a custom framework in AWS Audit Manager\cr
\link[=auditmanager_update_assessment_status]{update_assessment_status} \tab Updates the status of an assessment in AWS Audit Manager\cr
\link[=auditmanager_update_control]{update_control} \tab Updates a custom control in AWS Audit Manager\cr
\link[=auditmanager_update_settings]{update_settings} \tab Updates AWS Audit Manager settings for the current user account\cr
\link[=auditmanager_validate_assessment_report_integrity]{validate_assessment_report_integrity} \tab Validates the integrity of an assessment report in AWS Audit Manager
}
}

\examples{
\dontrun{
svc <- auditmanager()
svc$associate_assessment_report_evidence_folder(
  Foo = 123
)
}

}
