% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{backupgateway}
\alias{backupgateway}
\title{AWS Backup Gateway}
\usage{
backupgateway(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Backup gateway

Backup gateway connects Backup to your hypervisor, so you can create,
store, and restore backups of your virtual machines (VMs) anywhere,
whether on-premises or in the VMware Cloud (VMC) on Amazon Web Services.

Add on-premises resources by connecting to a hypervisor through a
gateway. Backup will automatically discover the resources in your
hypervisor.

Use Backup to assign virtual or on-premises resources to a backup plan,
or run on-demand backups. Once you have backed up your resources, you
can view them and restore them like any resource supported by Backup.

To download the Amazon Web Services software to get started, navigate to
the Backup console, choose \strong{Gateways}, then choose \strong{Create gateway}.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- backupgateway(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.networking:backupgateway_associate_gateway_to_server]{associate_gateway_to_server} \tab Associates a backup gateway with your server\cr
\link[paws.networking:backupgateway_create_gateway]{create_gateway} \tab Creates a backup gateway\cr
\link[paws.networking:backupgateway_delete_gateway]{delete_gateway} \tab Deletes a backup gateway\cr
\link[paws.networking:backupgateway_delete_hypervisor]{delete_hypervisor} \tab Deletes a hypervisor\cr
\link[paws.networking:backupgateway_disassociate_gateway_from_server]{disassociate_gateway_from_server} \tab Disassociates a backup gateway from the specified server\cr
\link[paws.networking:backupgateway_get_bandwidth_rate_limit_schedule]{get_bandwidth_rate_limit_schedule} \tab Retrieves the bandwidth rate limit schedule for a specified gateway\cr
\link[paws.networking:backupgateway_get_gateway]{get_gateway} \tab By providing the ARN (Amazon Resource Name), this API returns the gateway\cr
\link[paws.networking:backupgateway_get_hypervisor]{get_hypervisor} \tab This action requests information about the specified hypervisor to which the gateway will connect\cr
\link[paws.networking:backupgateway_get_hypervisor_property_mappings]{get_hypervisor_property_mappings} \tab This action retrieves the property mappings for the specified hypervisor\cr
\link[paws.networking:backupgateway_get_virtual_machine]{get_virtual_machine} \tab By providing the ARN (Amazon Resource Name), this API returns the virtual machine\cr
\link[paws.networking:backupgateway_import_hypervisor_configuration]{import_hypervisor_configuration} \tab Connect to a hypervisor by importing its configuration\cr
\link[paws.networking:backupgateway_list_gateways]{list_gateways} \tab Lists backup gateways owned by an Amazon Web Services account in an Amazon Web Services Region\cr
\link[paws.networking:backupgateway_list_hypervisors]{list_hypervisors} \tab Lists your hypervisors\cr
\link[paws.networking:backupgateway_list_tags_for_resource]{list_tags_for_resource} \tab Lists the tags applied to the resource identified by its Amazon Resource Name (ARN)\cr
\link[paws.networking:backupgateway_list_virtual_machines]{list_virtual_machines} \tab Lists your virtual machines\cr
\link[paws.networking:backupgateway_put_bandwidth_rate_limit_schedule]{put_bandwidth_rate_limit_schedule} \tab This action sets the bandwidth rate limit schedule for a specified gateway\cr
\link[paws.networking:backupgateway_put_hypervisor_property_mappings]{put_hypervisor_property_mappings} \tab This action sets the property mappings for the specified hypervisor\cr
\link[paws.networking:backupgateway_put_maintenance_start_time]{put_maintenance_start_time} \tab Set the maintenance start time for a gateway\cr
\link[paws.networking:backupgateway_start_virtual_machines_metadata_sync]{start_virtual_machines_metadata_sync} \tab This action sends a request to sync metadata across the specified virtual machines\cr
\link[paws.networking:backupgateway_tag_resource]{tag_resource} \tab Tag the resource\cr
\link[paws.networking:backupgateway_test_hypervisor_configuration]{test_hypervisor_configuration} \tab Tests your hypervisor configuration to validate that backup gateway can connect with the hypervisor and its resources\cr
\link[paws.networking:backupgateway_untag_resource]{untag_resource} \tab Removes tags from the resource\cr
\link[paws.networking:backupgateway_update_gateway_information]{update_gateway_information} \tab Updates a gateway's name\cr
\link[paws.networking:backupgateway_update_gateway_software_now]{update_gateway_software_now} \tab Updates the gateway virtual machine (VM) software\cr
\link[paws.networking:backupgateway_update_hypervisor]{update_hypervisor} \tab Updates a hypervisor metadata, including its host, username, and password
}
}

\examples{
\dontrun{
svc <- backupgateway()
svc$associate_gateway_to_server(
  Foo = 123
)
}

}
