% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_describe_jobs}
\alias{batch_describe_jobs}
\title{Describes a list of AWS Batch jobs}
\usage{
batch_describe_jobs(jobs)
}
\arguments{
\item{jobs}{[required] A space-separated list of up to 100 job IDs.}
}
\description{
Describes a list of AWS Batch jobs.
}
\section{Request syntax}{
\preformatted{batch$describe_jobs(
  jobs = list(
    "string"
  )
)
}
}

\examples{
# This example describes a job with the specified job ID.
\donttest{batch$describe_jobs(
  jobs = list(
    "24fa2d7a-64c4-49d2-8b47-f8da4fbde8e9"
  )
)}

}
\keyword{internal}
