% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{bedrockagentruntime}
\alias{bedrockagentruntime}
\title{Agents for Amazon Bedrock Runtime}
\usage{
bedrockagentruntime(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Contains APIs related to model invocation and querying of knowledge
bases.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- bedrockagentruntime(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.machine.learning:bedrockagentruntime_create_invocation]{create_invocation} \tab Creates a new invocation within a session\cr
\link[paws.machine.learning:bedrockagentruntime_create_session]{create_session} \tab Creates a session to temporarily store conversations for generative AI (GenAI) applications built with open-source frameworks such as LangGraph and LlamaIndex\cr
\link[paws.machine.learning:bedrockagentruntime_delete_agent_memory]{delete_agent_memory} \tab Deletes memory from the specified memory identifier\cr
\link[paws.machine.learning:bedrockagentruntime_delete_session]{delete_session} \tab Deletes a session that you ended\cr
\link[paws.machine.learning:bedrockagentruntime_end_session]{end_session} \tab Ends the session\cr
\link[paws.machine.learning:bedrockagentruntime_generate_query]{generate_query} \tab Generates an SQL query from a natural language query\cr
\link[paws.machine.learning:bedrockagentruntime_get_agent_memory]{get_agent_memory} \tab Gets the sessions stored in the memory of the agent\cr
\link[paws.machine.learning:bedrockagentruntime_get_execution_flow_snapshot]{get_execution_flow_snapshot} \tab Retrieves the flow definition snapshot used for a flow execution\cr
\link[paws.machine.learning:bedrockagentruntime_get_flow_execution]{get_flow_execution} \tab Retrieves details about a specific flow execution, including its status, start and end times, and any errors that occurred during execution\cr
\link[paws.machine.learning:bedrockagentruntime_get_invocation_step]{get_invocation_step} \tab Retrieves the details of a specific invocation step within an invocation in a session\cr
\link[paws.machine.learning:bedrockagentruntime_get_session]{get_session} \tab Retrieves details about a specific session\cr
\link[paws.machine.learning:bedrockagentruntime_invoke_agent]{invoke_agent} \tab Sends a prompt for the agent to process and respond to\cr
\link[paws.machine.learning:bedrockagentruntime_invoke_flow]{invoke_flow} \tab Invokes an alias of a flow to run the inputs that you specify and return the output of each node as a stream\cr
\link[paws.machine.learning:bedrockagentruntime_invoke_inline_agent]{invoke_inline_agent} \tab Invokes an inline Amazon Bedrock agent using the configurations you provide with the request\cr
\link[paws.machine.learning:bedrockagentruntime_list_flow_execution_events]{list_flow_execution_events} \tab Lists events that occurred during a flow execution\cr
\link[paws.machine.learning:bedrockagentruntime_list_flow_executions]{list_flow_executions} \tab Lists all executions of a flow\cr
\link[paws.machine.learning:bedrockagentruntime_list_invocations]{list_invocations} \tab Lists all invocations associated with a specific session\cr
\link[paws.machine.learning:bedrockagentruntime_list_invocation_steps]{list_invocation_steps} \tab Lists all invocation steps associated with a session and optionally, an invocation within the session\cr
\link[paws.machine.learning:bedrockagentruntime_list_sessions]{list_sessions} \tab Lists all sessions in your Amazon Web Services account\cr
\link[paws.machine.learning:bedrockagentruntime_list_tags_for_resource]{list_tags_for_resource} \tab List all the tags for the resource you specify\cr
\link[paws.machine.learning:bedrockagentruntime_optimize_prompt]{optimize_prompt} \tab Optimizes a prompt for the task that you specify\cr
\link[paws.machine.learning:bedrockagentruntime_put_invocation_step]{put_invocation_step} \tab Add an invocation step to an invocation in a session\cr
\link[paws.machine.learning:bedrockagentruntime_rerank]{rerank} \tab Reranks the relevance of sources based on queries\cr
\link[paws.machine.learning:bedrockagentruntime_retrieve]{retrieve} \tab Queries a knowledge base and retrieves information from it\cr
\link[paws.machine.learning:bedrockagentruntime_retrieve_and_generate]{retrieve_and_generate} \tab Queries a knowledge base and generates responses based on the retrieved results and using the specified foundation model or inference profile\cr
\link[paws.machine.learning:bedrockagentruntime_retrieve_and_generate_stream]{retrieve_and_generate_stream} \tab Queries a knowledge base and generates responses based on the retrieved results, with output in streaming format\cr
\link[paws.machine.learning:bedrockagentruntime_start_flow_execution]{start_flow_execution} \tab Starts an execution of an Amazon Bedrock flow\cr
\link[paws.machine.learning:bedrockagentruntime_stop_flow_execution]{stop_flow_execution} \tab Stops an Amazon Bedrock flow's execution\cr
\link[paws.machine.learning:bedrockagentruntime_tag_resource]{tag_resource} \tab Associate tags with a resource\cr
\link[paws.machine.learning:bedrockagentruntime_untag_resource]{untag_resource} \tab Remove tags from a resource\cr
\link[paws.machine.learning:bedrockagentruntime_update_session]{update_session} \tab Updates the metadata or encryption settings of a session
}
}

\examples{
\dontrun{
svc <- bedrockagentruntime()
svc$create_invocation(
  Foo = 123
)
}

}
