% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_service.R
\name{braket}
\alias{braket}
\title{Braket}
\usage{
braket(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
The Amazon Braket API Reference provides information about the
operations and structures supported in Amazon Braket.
}
\section{Service syntax}{
\preformatted{svc <- braket(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=braket_cancel_quantum_task]{cancel_quantum_task} \tab Cancels the specified task\cr
\link[=braket_create_quantum_task]{create_quantum_task} \tab Creates a quantum task\cr
\link[=braket_get_device]{get_device} \tab Retrieves the devices available in Amazon Braket\cr
\link[=braket_get_quantum_task]{get_quantum_task} \tab Retrieves the specified quantum task\cr
\link[=braket_list_tags_for_resource]{list_tags_for_resource} \tab Shows the tags associated with this resource\cr
\link[=braket_search_devices]{search_devices} \tab Searches for devices using the specified filters\cr
\link[=braket_search_quantum_tasks]{search_quantum_tasks} \tab Searches for tasks that match the specified filter values\cr
\link[=braket_tag_resource]{tag_resource} \tab Add a tag to the specified resource\cr
\link[=braket_untag_resource]{untag_resource} \tab Remove tags from a resource
}
}

\examples{
\dontrun{
svc <- braket()
svc$cancel_quantum_task(
  Foo = 123
)
}

}
