% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud9_service.R
\name{cloud9}
\alias{cloud9}
\title{AWS Cloud9}
\usage{
cloud9(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
AWS Cloud9 is a collection of tools that you can use to code, build,
run, test, debug, and release software in the cloud.

For more information about AWS Cloud9, see the \href{https://docs.aws.amazon.com/cloud9/latest/user-guide}{AWS Cloud9 User Guide}.

AWS Cloud9 supports these operations:
\itemize{
\item \code{CreateEnvironmentEC2}: Creates an AWS Cloud9 development
environment, launches an Amazon EC2 instance, and then connects from
the instance to the environment.
\item \code{CreateEnvironmentMembership}: Adds an environment member to an
environment.
\item \code{DeleteEnvironment}: Deletes an environment. If an Amazon EC2
instance is connected to the environment, also terminates the
instance.
\item \code{DeleteEnvironmentMembership}: Deletes an environment member from an
environment.
\item \code{DescribeEnvironmentMemberships}: Gets information about environment
members for an environment.
\item \code{DescribeEnvironments}: Gets information about environments.
\item \code{DescribeEnvironmentStatus}: Gets status information for an
environment.
\item \code{ListEnvironments}: Gets a list of environment identifiers.
\item \code{UpdateEnvironment}: Changes the settings of an existing
environment.
\item \code{UpdateEnvironmentMembership}: Changes the settings of an existing
environment member for an environment.
}
}
\section{Service syntax}{
\preformatted{svc <- cloud9(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=cloud9_create_environment_ec2]{create_environment_ec2} \tab Creates an AWS Cloud9 development environment, launches an Amazon Elastic Compute Cloud (Amazon EC2) instance, and then connects from the instance to the environment\cr
\link[=cloud9_create_environment_membership]{create_environment_membership} \tab Adds an environment member to an AWS Cloud9 development environment \cr
\link[=cloud9_delete_environment]{delete_environment} \tab Deletes an AWS Cloud9 development environment \cr
\link[=cloud9_delete_environment_membership]{delete_environment_membership} \tab Deletes an environment member from an AWS Cloud9 development environment \cr
\link[=cloud9_describe_environment_memberships]{describe_environment_memberships} \tab Gets information about environment members for an AWS Cloud9 development environment \cr
\link[=cloud9_describe_environments]{describe_environments} \tab Gets information about AWS Cloud9 development environments \cr
\link[=cloud9_describe_environment_status]{describe_environment_status} \tab Gets status information for an AWS Cloud9 development environment \cr
\link[=cloud9_list_environments]{list_environments} \tab Gets a list of AWS Cloud9 development environment identifiers \cr
\link[=cloud9_update_environment]{update_environment} \tab Changes the settings of an existing AWS Cloud9 development environment \cr
\link[=cloud9_update_environment_membership]{update_environment_membership} \tab Changes the settings of an existing environment member for an AWS Cloud9 development environment
}
}

\examples{
# 
\donttest{svc <- cloud9()
svc$create_environment_ec2(
  name = "my-demo-environment",
  automaticStopTimeMinutes = 60L,
  description = "This is my demonstration environment.",
  instanceType = "t2.micro",
  ownerArn = "arn:aws:iam::123456789012:user/MyDemoUser",
  subnetId = "subnet-1fab8aEX"
)}

}
