% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{cloudwatchobservabilityaccessmanager}
\alias{cloudwatchobservabilityaccessmanager}
\title{CloudWatch Observability Access Manager}
\usage{
cloudwatchobservabilityaccessmanager(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Use Amazon CloudWatch Observability Access Manager to create and manage
links between source accounts and monitoring accounts by using
\emph{CloudWatch cross-account observability}. With CloudWatch cross-account
observability, you can monitor and troubleshoot applications that span
multiple accounts within a Region. Seamlessly search, visualize, and
analyze your metrics, logs, traces, Application Signals services and
service level objectives (SLOs), Application Insights applications, and
internet monitors in any of the linked accounts without account
boundaries.

Set up one or more Amazon Web Services accounts as \emph{monitoring accounts}
and link them with multiple \emph{source accounts}. A monitoring account is a
central Amazon Web Services account that can view and interact with
observability data generated from source accounts. A source account is
an individual Amazon Web Services account that generates observability
data for the resources that reside in it. Source accounts share their
observability data with the monitoring account. The shared observability
data can include metrics in Amazon CloudWatch, logs in Amazon CloudWatch
Logs, traces in X-Ray, Application Signals services and service level
objectives (SLOs), applications in Amazon CloudWatch Application
Insights, and internet monitors in CloudWatch Internet Monitor.

When you set up a link, you can choose to share the metrics from all
namespaces with the monitoring account, or filter to a subset of
namespaces. And for CloudWatch Logs, you can choose to share all log
groups with the monitoring account, or filter to a subset of log groups.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- cloudwatchobservabilityaccessmanager(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.management:cloudwatchobservabilityaccessmanager_create_link]{create_link} \tab Creates a link between a source account and a sink that you have created in a monitoring account\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_create_sink]{create_sink} \tab Use this to create a sink in the current account, so that it can be used as a monitoring account in CloudWatch cross-account observability\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_delete_link]{delete_link} \tab Deletes a link between a monitoring account sink and a source account\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_delete_sink]{delete_sink} \tab Deletes a sink\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_get_link]{get_link} \tab Returns complete information about one link\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_get_sink]{get_sink} \tab Returns complete information about one monitoring account sink\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_get_sink_policy]{get_sink_policy} \tab Returns the current sink policy attached to this sink\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_list_attached_links]{list_attached_links} \tab Returns a list of source account links that are linked to this monitoring account sink\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_list_links]{list_links} \tab Use this operation in a source account to return a list of links to monitoring account sinks that this source account has\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_list_sinks]{list_sinks} \tab Use this operation in a monitoring account to return the list of sinks created in that account\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_list_tags_for_resource]{list_tags_for_resource} \tab Displays the tags associated with a resource\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_put_sink_policy]{put_sink_policy} \tab Creates or updates the resource policy that grants permissions to source accounts to link to the monitoring account sink\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_tag_resource]{tag_resource} \tab Assigns one or more tags (key-value pairs) to the specified resource\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_untag_resource]{untag_resource} \tab Removes one or more tags from the specified resource\cr
\link[paws.management:cloudwatchobservabilityaccessmanager_update_link]{update_link} \tab Use this operation to change what types of data are shared from a source account to its linked monitoring account sink
}
}

\examples{
\dontrun{
svc <- cloudwatchobservabilityaccessmanager()
svc$create_link(
  Foo = 123
)
}

}
