% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_dispose_package_versions}
\alias{codeartifact_dispose_package_versions}
\title{Deletes the assets in package versions and sets the package versions'
status to Disposed}
\usage{
codeartifact_dispose_package_versions(domain, domainOwner, repository,
  format, namespace, package, versions, versionRevisions, expectedStatus)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the repository you want to dispose.}

\item{domainOwner}{The 12-digit account number of the AWS account that owns the domain. It
does not include dashes or spaces.}

\item{repository}{[required] The name of the repository that contains the package versions you want
to dispose.}

\item{format}{[required] A format that specifies the type of package versions you want to
dispose. The valid values are:
\itemize{
\item \code{npm}
\item \code{pypi}
\item \code{maven}
}}

\item{namespace}{The namespace of the package. The package component that specifies its
namespace depends on its type. For example:
\itemize{
\item The namespace of a Maven package is its \code{groupId}.
\item The namespace of an npm package is its \code{scope}.
\item A Python package does not contain a corresponding component, so
Python packages do not have a namespace.
}}

\item{package}{[required] The name of the package with the versions you want to dispose.}

\item{versions}{[required] The versions of the package you want to dispose.}

\item{versionRevisions}{The revisions of the package versions you want to dispose.}

\item{expectedStatus}{The expected status of the package version to dispose. Valid values are:
\itemize{
\item \code{Published}
\item \code{Unfinished}
\item \code{Unlisted}
\item \code{Archived}
\item \code{Disposed}
}}
}
\description{
Deletes the assets in package versions and sets the package versions\'
status to \code{Disposed}. A disposed package version cannot be restored in
your repository because its assets are deleted.
}
\details{
To view all disposed package versions in a repository, use
\verb{<a href="https://docs.aws.amazon.com/codeartifact/latest/APIReference/API_ListPackageVersions.html">ListackageVersions</a>}
and set the
\verb{<a href="https://docs.aws.amazon.com/codeartifact/latest/APIReference/API_ListPackageVersions.html#API_ListPackageVersions_RequestSyntax">status</a>}
parameter to \code{Disposed}.

To view information about a disposed package version, use
\verb{<a href="https://docs.aws.amazon.com/codeartifact/latest/APIReference/API_ListPackageVersions.html">ListPackageVersions</a>}
and set the
\verb{<a href="https://docs.aws.amazon.com/API_ListPackageVersions.html#codeartifact-ListPackageVersions-response-status">status</a>}
parameter to \code{Disposed}.
}
\section{Request syntax}{
\preformatted{svc$dispose_package_versions(
  domain = "string",
  domainOwner = "string",
  repository = "string",
  format = "npm"|"pypi"|"maven",
  namespace = "string",
  package = "string",
  versions = list(
    "string"
  ),
  versionRevisions = list(
    "string"
  ),
  expectedStatus = "Published"|"Unfinished"|"Unlisted"|"Archived"|"Disposed"|"Deleted"
)
}
}

\keyword{internal}
