% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_batch_get_deployments}
\alias{codedeploy_batch_get_deployments}
\title{Gets information about one or more deployments}
\usage{
codedeploy_batch_get_deployments(deploymentIds)
}
\arguments{
\item{deploymentIds}{[required] A list of deployment IDs, separated by spaces.}
}
\description{
Gets information about one or more deployments.
}
\section{Request syntax}{
\preformatted{codedeploy$batch_get_deployments(
  deploymentIds = list(
    "string"
  )
)
}
}

\keyword{internal}
