% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_post_agent_profile}
\alias{codeguruprofiler_post_agent_profile}
\title{Submit profile collected by an agent belonging to a profiling group for
aggregation}
\usage{
codeguruprofiler_post_agent_profile(agentProfile, contentType,
  profileToken, profilingGroupName)
}
\arguments{
\item{agentProfile}{[required]}

\item{contentType}{[required] The content type of the agent profile in the payload. Recommended to
send the profile gzipped with content-type application/octet-stream.
Other accepted values are application/x-amzn-ion and application/json
for unzipped Ion and JSON respectively.}

\item{profileToken}{Client generated token to deduplicate the agent profile during
aggregation.}

\item{profilingGroupName}{[required]}
}
\description{
Submit profile collected by an agent belonging to a profiling group for
aggregation.
}
\section{Request syntax}{
\preformatted{svc$post_agent_profile(
  agentProfile = raw,
  contentType = "string",
  profileToken = "string",
  profilingGroupName = "string"
)
}
}

\keyword{internal}
