% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_put_permission}
\alias{codeguruprofiler_put_permission}
\title{Adds permissions to a profiling group's resource-based policy that are
provided using an action group}
\usage{
codeguruprofiler_put_permission(actionGroup, principals,
  profilingGroupName, revisionId)
}
\arguments{
\item{actionGroup}{[required] Specifies an action group that contains permissions to add to a
profiling group resource. One action group is supported,
\code{agentPermissions}, which grants permission to perform actions required
by the profiling agent, \code{ConfigureAgent} and \code{PostAgentProfile}
permissions.}

\item{principals}{[required] A list ARNs for the roles and users you want to grant access to the
profiling group. Wildcards are not are supported in the ARNs.}

\item{profilingGroupName}{[required] The name of the profiling group to grant access to.}

\item{revisionId}{A universally unique identifier (UUID) for the revision of the policy
you are adding to the profiling group. Do not specify this when you add
permissions to a profiling group for the first time. If a policy already
exists on the profiling group, you must specify the \code{revisionId}.}
}
\description{
Adds permissions to a profiling group's resource-based policy that are
provided using an action group. If a profiling group doesn't have a
resource-based policy, one is created for it using the permissions in
the action group and the roles and users in the \code{principals} parameter.
}
\details{
\preformatted{ &lt;p&gt; The one supported action group that can be added is &lt;code&gt;agentPermission&lt;/code&gt; which grants &lt;code&gt;ConfigureAgent&lt;/code&gt; and &lt;code&gt;PostAgent&lt;/code&gt; permissions. For more information, see &lt;a href="https://docs.aws.amazon.com/codeguru/latest/profiler-ug/resource-based-policies.html"&gt;Resource-based policies in CodeGuru Profiler&lt;/a&gt; in the &lt;i&gt;Amazon CodeGuru Profiler User Guide&lt;/i&gt;, &lt;a href="https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_ConfigureAgent.html"&gt; &lt;code&gt;ConfigureAgent&lt;/code&gt; &lt;/a&gt;, and &lt;a href="https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_PostAgentProfile.html"&gt; &lt;code&gt;PostAgentProfile&lt;/code&gt; &lt;/a&gt;. &lt;/p&gt; &lt;p&gt; The first time you call &lt;code&gt;PutPermission&lt;/code&gt; on a profiling group, do not specify a &lt;code&gt;revisionId&lt;/code&gt; because it doesn't have a resource-based policy. Subsequent calls must provide a &lt;code&gt;revisionId&lt;/code&gt; to specify which revision of the resource-based policy to add the permissions to. &lt;/p&gt; &lt;p&gt; The response contains the profiling group's JSON-formatted resource policy. &lt;/p&gt;
}
}
\section{Request syntax}{
\preformatted{svc$put_permission(
  actionGroup = "agentPermissions",
  principals = list(
    "string"
  ),
  profilingGroupName = "string",
  revisionId = "string"
)
}
}

\keyword{internal}
