% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_update_profiling_group}
\alias{codeguruprofiler_update_profiling_group}
\title{Update a profiling group}
\usage{
codeguruprofiler_update_profiling_group(agentOrchestrationConfig,
  profilingGroupName)
}
\arguments{
\item{agentOrchestrationConfig}{[required] Remote configuration to configure the agents of the profiling group.}

\item{profilingGroupName}{[required]}
}
\description{
Update a profiling group.
}
\section{Request syntax}{
\preformatted{svc$update_profiling_group(
  agentOrchestrationConfig = list(
    profilingEnabled = TRUE|FALSE
  ),
  profilingGroupName = "string"
)
}
}

\keyword{internal}
