% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_service.R
\name{codegurureviewer}
\alias{codegurureviewer}
\title{Amazon CodeGuru Reviewer}
\usage{
codegurureviewer(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
This section provides documentation for the Amazon CodeGuru Reviewer API
operations.
}
\section{Service syntax}{
\preformatted{svc <- codegurureviewer(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=codegurureviewer_associate_repository]{associate_repository} \tab Associates an AWS CodeCommit repository with Amazon CodeGuru Reviewer \cr
\link[=codegurureviewer_describe_repository_association]{describe_repository_association} \tab Describes a repository association \cr
\link[=codegurureviewer_disassociate_repository]{disassociate_repository} \tab Removes the association between Amazon CodeGuru Reviewer and a repository\cr
\link[=codegurureviewer_list_repository_associations]{list_repository_associations} \tab Lists repository associations
}
}

\examples{
\donttest{svc <- codegurureviewer()
svc$associate_repository(
  Foo = 123
)}

}
