% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_associate_repository}
\alias{codegurureviewer_associate_repository}
\title{Use to associate an AWS CodeCommit repository or a repostory managed by
AWS CodeStar Connections with Amazon CodeGuru Reviewer}
\usage{
codegurureviewer_associate_repository(Repository, ClientRequestToken)
}
\arguments{
\item{Repository}{[required] The repository to associate.}

\item{ClientRequestToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.

To add a new repository association, this parameter specifies a unique
identifier for the new repository association that helps ensure
idempotency.

If you use the AWS CLI or one of the AWS SDKs to call this operation,
you can leave this parameter empty. The CLI or SDK generates a random
UUID for you and includes that in the request. If you don\'t use the SDK
and instead generate a raw HTTP request to the Secrets Manager service
endpoint, you must generate a ClientRequestToken yourself for new
versions and include that value in the request.

You typically interact with this value if you implement your own retry
logic and want to ensure that a given repository association is not
created twice. We recommend that you generate a UUID-type value to
ensure uniqueness within the specified repository association.

Amazon CodeGuru Reviewer uses this value to prevent the accidental
creation of duplicate repository associations if there are failures and
retries.}
}
\description{
Use to associate an AWS CodeCommit repository or a repostory managed by
AWS CodeStar Connections with Amazon CodeGuru Reviewer. When you
associate a repository, CodeGuru Reviewer reviews source code changes in
the repository\'s pull requests and provides automatic recommendations.
You can view recommendations using the CodeGuru Reviewer console. For
more information, see \href{https://docs.aws.amazon.com/codeguru/latest/reviewer-ug/recommendations.html}{Recommendations in Amazon CodeGuru Reviewer}
in the \emph{Amazon CodeGuru Reviewer User Guide.}
}
\details{
If you associate a CodeCommit repository, it must be in the same AWS
Region and AWS account where its CodeGuru Reviewer code reviews are
configured.

Bitbucket and GitHub Enterprise Server repositories are managed by AWS
CodeStar Connections to connect to CodeGuru Reviewer. For more
information, see Connect to a repository source provider in the \emph{Amazon
CodeGuru Reviewer User Guide.}

You cannot use the CodeGuru Reviewer SDK or the AWS CLI to associate a
GitHub repository with Amazon CodeGuru Reviewer. To associate a GitHub
repository, use the console. For more information, see \href{https://docs.aws.amazon.com/codeguru/latest/reviewer-ug/getting-started-with-guru.html}{Getting started with CodeGuru Reviewer}
in the \emph{CodeGuru Reviewer User Guide.}
}
\section{Request syntax}{
\preformatted{svc$associate_repository(
  Repository = list(
    CodeCommit = list(
      Name = "string"
    ),
    Bitbucket = list(
      Name = "string",
      ConnectionArn = "string",
      Owner = "string"
    ),
    GitHubEnterpriseServer = list(
      Name = "string",
      ConnectionArn = "string",
      Owner = "string"
    )
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
