% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_associate_repository}
\alias{codegurureviewer_associate_repository}
\title{Use to associate an AWS CodeCommit repository or a repostory managed by
AWS CodeStar Connections with Amazon CodeGuru Reviewer}
\usage{
codegurureviewer_associate_repository(Repository, ClientRequestToken,
  Tags)
}
\arguments{
\item{Repository}{[required] The repository to associate.}

\item{ClientRequestToken}{Amazon CodeGuru Reviewer uses this value to prevent the accidental
creation of duplicate repository associations if there are failures and
retries.}

\item{Tags}{An array of key-value pairs used to tag an associated repository. A tag
is a custom attribute label with two parts:
\itemize{
\item A \emph{tag key} (for example, \code{CostCenter}, \code{Environment}, \code{Project}, or
\code{Secret}). Tag keys are case sensitive.
\item An optional field known as a \emph{tag value} (for example,
\code{111122223333}, \code{Production}, or a team name). Omitting the tag
value is the same as using an empty string. Like tag keys, tag
values are case sensitive.
}}
}
\description{
Use to associate an AWS CodeCommit repository or a repostory managed by
AWS CodeStar Connections with Amazon CodeGuru Reviewer. When you
associate a repository, CodeGuru Reviewer reviews source code changes in
the repository's pull requests and provides automatic recommendations.
You can view recommendations using the CodeGuru Reviewer console. For
more information, see \href{https://docs.aws.amazon.com/codeguru/latest/reviewer-ug/recommendations.html}{Recommendations in Amazon CodeGuru Reviewer}
in the \emph{Amazon CodeGuru Reviewer User Guide.}

If you associate a CodeCommit repository, it must be in the same AWS
Region and AWS account where its CodeGuru Reviewer code reviews are
configured.

Bitbucket and GitHub Enterprise Server repositories are managed by AWS
CodeStar Connections to connect to CodeGuru Reviewer. For more
information, see Connect to a repository source provider in the \emph{Amazon
CodeGuru Reviewer User Guide.}

You cannot use the CodeGuru Reviewer SDK or the AWS CLI to associate a
GitHub repository with Amazon CodeGuru Reviewer. To associate a GitHub
repository, use the console. For more information, see \href{https://docs.aws.amazon.com/codeguru/latest/reviewer-ug/getting-started-with-guru.html}{Getting started with CodeGuru Reviewer}
in the \emph{CodeGuru Reviewer User Guide.}
}
\section{Request syntax}{
\preformatted{svc$associate_repository(
  Repository = list(
    CodeCommit = list(
      Name = "string"
    ),
    Bitbucket = list(
      Name = "string",
      ConnectionArn = "string",
      Owner = "string"
    ),
    GitHubEnterpriseServer = list(
      Name = "string",
      ConnectionArn = "string",
      Owner = "string"
    )
  ),
  ClientRequestToken = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
