% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_service.R
\name{codestarconnections}
\alias{codestarconnections}
\title{AWS CodeStar connections}
\usage{
codestarconnections(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
This AWS CodeStar Connections API Reference provides descriptions and
usage examples of the operations and data types for the AWS CodeStar
Connections API. You can use the Connections API to work with
connections and installations.

\emph{Connections} are configurations that you use to connect AWS resources
to external code repositories. Each connection is a resource that can be
given to services such as CodePipeline to connect to a third-party
repository such as Bitbucket. For example, you can add the connection in
CodePipeline so that it triggers your pipeline when a code change is
made to your third-party code repository. Each connection is named and
associated with a unique ARN that is used to reference the connection.

When you create a connection, the console initiates a third-party
connection handshake. \emph{Installations} are the apps that are used to
conduct this handshake. For example, the installation for the Bitbucket
provider type is the Bitbucket Cloud app. When you create a connection,
you can choose an existing installation or create one.

You can work with connections by calling:
\itemize{
\item CreateConnection, which creates a uniquely named connection that can
be referenced by services such as CodePipeline.
\item DeleteConnection, which deletes the specified connection.
\item GetConnection, which returns information about the connection,
including the connection status.
\item ListConnections, which lists the connections associated with your
account.
}

For information about how to use AWS CodeStar Connections, see the \href{https://docs.aws.amazon.com/codepipeline/latest/userguide/welcome.html}{AWS CodePipeline User Guide}.
}
\section{Service syntax}{
\preformatted{svc <- codestarconnections(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Examples}{
\preformatted{svc$create_connection(
  Foo = 123
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=codestarconnections_create_connection]{create_connection} \tab Creates a connection that can then be given to other AWS services like CodePipeline so that it can access third-party code repositories\cr
\link[=codestarconnections_delete_connection]{delete_connection} \tab The connection to be deleted \cr
\link[=codestarconnections_get_connection]{get_connection} \tab Returns the connection ARN and details such as status, owner, and provider type \cr
\link[=codestarconnections_list_connections]{list_connections} \tab Lists the connections associated with your account
}
}

