% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_service.R
\name{codestarnotifications}
\alias{codestarnotifications}
\title{AWS CodeStar Notifications}
\usage{
codestarnotifications(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
This AWS CodeStar Notifications API Reference provides descriptions and
usage examples of the operations and data types for the AWS CodeStar
Notifications API. You can use the AWS CodeStar Notifications API to
work with the following objects:

Notification rules, by calling the following:
\itemize{
\item CreateNotificationRule, which creates a notification rule for a
resource in your account.
\item DeleteNotificationRule, which deletes a notification rule.
\item DescribeNotificationRule, which provides information about a
notification rule.
\item ListNotificationRules, which lists the notification rules associated
with your account.
\item UpdateNotificationRule, which changes the name, events, or targets
associated with a notification rule.
\item Subscribe, which subscribes a target to a notification rule.
\item Unsubscribe, which removes a target from a notification rule.
}

Targets, by calling the following:
\itemize{
\item DeleteTarget, which removes a notification rule target (SNS topic)
from a notification rule.
\item ListTargets, which lists the targets associated with a notification
rule.
}

Events, by calling the following:
\itemize{
\item ListEventTypes, which lists the event types you can include in a
notification rule.
}

Tags, by calling the following:
\itemize{
\item ListTagsForResource, which lists the tags already associated with a
notification rule in your account.
\item TagResource, which associates a tag you provide with a notification
rule in your account.
\item UntagResource, which removes a tag from a notification rule in your
account.
}

For information about how to use AWS CodeStar Notifications, see link in
the CodeStarNotifications User Guide.
}
\section{Service syntax}{
\preformatted{svc <- codestarnotifications(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=codestarnotifications_create_notification_rule]{create_notification_rule} \tab Creates a notification rule for a resource \cr
\link[=codestarnotifications_delete_notification_rule]{delete_notification_rule} \tab Deletes a notification rule for a resource \cr
\link[=codestarnotifications_delete_target]{delete_target} \tab Deletes a specified target for notifications \cr
\link[=codestarnotifications_describe_notification_rule]{describe_notification_rule} \tab Returns information about a specified notification rule \cr
\link[=codestarnotifications_list_event_types]{list_event_types} \tab Returns information about the event types available for configuring notifications \cr
\link[=codestarnotifications_list_notification_rules]{list_notification_rules} \tab Returns a list of the notification rules for an AWS account \cr
\link[=codestarnotifications_list_tags_for_resource]{list_tags_for_resource} \tab Returns a list of the tags associated with a notification rule \cr
\link[=codestarnotifications_list_targets]{list_targets} \tab Returns a list of the notification rule targets for an AWS account \cr
\link[=codestarnotifications_subscribe]{subscribe} \tab Creates an association between a notification rule and an SNS topic so that the associated target can receive notifications when the events described in the rule are triggered \cr
\link[=codestarnotifications_tag_resource]{tag_resource} \tab Associates a set of provided tags with a notification rule \cr
\link[=codestarnotifications_unsubscribe]{unsubscribe} \tab Removes an association between a notification rule and an Amazon SNS topic so that subscribers to that topic stop receiving notifications when the events described in the rule are triggered\cr
\link[=codestarnotifications_untag_resource]{untag_resource} \tab Removes the association between one or more provided tags and a notification rule \cr
\link[=codestarnotifications_update_notification_rule]{update_notification_rule} \tab Updates a notification rule for a resource
}
}

\examples{
\dontrun{
svc <- codestarnotifications()
svc$create_notification_rule(
  Foo = 123
)
}

}
