% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_service.R
\name{cognitoidentity}
\alias{cognitoidentity}
\title{Amazon Cognito Identity}
\usage{
cognitoidentity()
}
\description{
Amazon Cognito

Amazon Cognito is a web service that delivers scoped temporary
credentials to mobile devices and other untrusted environments. Amazon
Cognito uniquely identifies a device and supplies the user with a
consistent identity over the lifetime of an application.

Using Amazon Cognito, you can enable authentication with one or more
third-party identity providers (Facebook, Google, or Login with Amazon),
and you can also choose to support unauthenticated access from your app.
Cognito delivers a unique identifier for each user and acts as an OpenID
token provider trusted by AWS Security Token Service (STS) to access
temporary, limited-privilege AWS credentials.

To provide end-user credentials, first make an unsigned call to GetId.
If the end user is authenticated with one of the supported identity
providers, set the \code{Logins} map with the identity provider token.
\code{GetId} returns a unique identifier for the user.

Next, make an unsigned call to GetCredentialsForIdentity. This call
expects the same \code{Logins} map as the \code{GetId} call, as well as the
\code{IdentityID} originally returned by \code{GetId}. Assuming your identity pool
has been configured via the SetIdentityPoolRoles operation,
\code{GetCredentialsForIdentity} will return AWS credentials for your use. If
your pool has not been configured with \code{SetIdentityPoolRoles}, or if you
want to follow legacy flow, make an unsigned call to GetOpenIdToken,
which returns the OpenID token necessary to call STS and retrieve AWS
credentials. This call expects the same \code{Logins} map as the \code{GetId}
call, as well as the \code{IdentityID} originally returned by \code{GetId}. The
token returned by \code{GetOpenIdToken} can be passed to the STS operation
\href{http://docs.aws.amazon.com/STS/latest/APIReference/API_AssumeRoleWithWebIdentity.html}{AssumeRoleWithWebIdentity}
to retrieve AWS credentials.

If you want to use Amazon Cognito in an Android, iOS, or Unity
application, you will probably want to make API calls via the AWS Mobile
SDK. To learn more, see the \href{http://docs.aws.amazon.com/mobile/index.html}{AWS Mobile SDK Developer Guide}.
}
\section{Example}{
\preformatted{cognitoidentity <- paws::cognitoidentity()
cognitoidentity$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=cognitoidentity_create_identity_pool]{create_identity_pool} \tab Creates a new identity pool \cr
\link[=cognitoidentity_delete_identities]{delete_identities} \tab Deletes identities from an identity pool \cr
\link[=cognitoidentity_delete_identity_pool]{delete_identity_pool} \tab Deletes a user pool \cr
\link[=cognitoidentity_describe_identity]{describe_identity} \tab Returns metadata related to the given identity, including when the identity was created and any associated linked logins \cr
\link[=cognitoidentity_describe_identity_pool]{describe_identity_pool} \tab Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users \cr
\link[=cognitoidentity_get_credentials_for_identity]{get_credentials_for_identity} \tab Returns credentials for the provided identity ID \cr
\link[=cognitoidentity_get_id]{get_id} \tab Generates (or retrieves) a Cognito ID \cr
\link[=cognitoidentity_get_identity_pool_roles]{get_identity_pool_roles} \tab Gets the roles for an identity pool \cr
\link[=cognitoidentity_get_open_id_token]{get_open_id_token} \tab Gets an OpenID token, using a known Cognito ID \cr
\link[=cognitoidentity_get_open_id_token_for_developer_identity]{get_open_id_token_for_developer_identity} \tab Registers (or retrieves) a Cognito IdentityId and an OpenID Connect token for a user authenticated by your backend authentication process \cr
\link[=cognitoidentity_list_identities]{list_identities} \tab Lists the identities in a pool \cr
\link[=cognitoidentity_list_identity_pools]{list_identity_pools} \tab Lists all of the Cognito identity pools registered for your account \cr
\link[=cognitoidentity_lookup_developer_identity]{lookup_developer_identity} \tab Retrieves the IdentityID associated with a DeveloperUserIdentifier or the list of DeveloperUserIdentifiers associated with an IdentityId for an existing identity\cr
\link[=cognitoidentity_merge_developer_identities]{merge_developer_identities} \tab Merges two users having different IdentityIds, existing in the same identity pool, and identified by the same developer provider \cr
\link[=cognitoidentity_set_identity_pool_roles]{set_identity_pool_roles} \tab Sets the roles for an identity pool \cr
\link[=cognitoidentity_unlink_developer_identity]{unlink_developer_identity} \tab Unlinks a DeveloperUserIdentifier from an existing identity \cr
\link[=cognitoidentity_unlink_identity]{unlink_identity} \tab Unlinks a federated identity from an existing account \cr
\link[=cognitoidentity_update_identity_pool]{update_identity_pool} \tab Updates a user pool
}
}

