% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_set_user_mfa_preference}
\alias{cognitoidentityprovider_set_user_mfa_preference}
\title{Set the user's multi-factor authentication (MFA) method preference}
\usage{
cognitoidentityprovider_set_user_mfa_preference(SMSMfaSettings,
  SoftwareTokenMfaSettings, AccessToken)
}
\arguments{
\item{SMSMfaSettings}{The SMS text message multi-factor authentication (MFA) settings.}

\item{SoftwareTokenMfaSettings}{The time-based one-time password software token MFA settings.}

\item{AccessToken}{[required] The access token.}
}
\description{
Set the user's multi-factor authentication (MFA) method preference.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$set_user_mfa_preference(
  SMSMfaSettings = list(
    Enabled = TRUE|FALSE,
    PreferredMfa = TRUE|FALSE
  ),
  SoftwareTokenMfaSettings = list(
    Enabled = TRUE|FALSE,
    PreferredMfa = TRUE|FALSE
  ),
  AccessToken = "string"
)
}
}

\keyword{internal}
