% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_start_document_classification_job}
\alias{comprehend_start_document_classification_job}
\title{Starts an asynchronous document classification job}
\usage{
comprehend_start_document_classification_job(JobName,
  DocumentClassifierArn, InputDataConfig, OutputDataConfig,
  DataAccessRoleArn, ClientRequestToken)
}
\arguments{
\item{JobName}{The identifier of the job.}

\item{DocumentClassifierArn}{[required] The Amazon Resource Name (ARN) of the document classifier to use to
process the job.}

\item{InputDataConfig}{[required] Specifies the format and location of the input data for the job.}

\item{OutputDataConfig}{[required] Specifies where to send the output files.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Access Management
(IAM) role that grants Amazon Comprehend read access to your input data.}

\item{ClientRequestToken}{A unique identifier for the request. If you do not set the client
request token, Amazon Comprehend generates one.}
}
\description{
Starts an asynchronous document classification job. Use the operation to
track the progress of the job.
}
\section{Request syntax}{
\preformatted{comprehend$start_document_classification_job(
  JobName = "string",
  DocumentClassifierArn = "string",
  InputDataConfig = list(
    S3Uri = "string",
    InputFormat = "ONE_DOC_PER_FILE"|"ONE_DOC_PER_LINE"
  ),
  OutputDataConfig = list(
    S3Uri = "string"
  ),
  DataAccessRoleArn = "string",
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
