% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_detect_entities}
\alias{comprehendmedical_detect_entities}
\title{Inspects the clinical text for a variety of medical entities and returns
specific information about them such as entity category, location, and
confidence score on that information}
\usage{
comprehendmedical_detect_entities(Text)
}
\arguments{
\item{Text}{[required] A UTF-8 text string containing the clinical content being examined for
entities. Each string must contain fewer than 20,000 bytes of
characters.}
}
\description{
Inspects the clinical text for a variety of medical entities and returns
specific information about them such as entity category, location, and
confidence score on that information .
}
\section{Request syntax}{
\preformatted{comprehendmedical$detect_entities(
  Text = "string"
)
}
}

\keyword{internal}
