% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_service.R
\name{computeoptimizer}
\alias{computeoptimizer}
\title{AWS Compute Optimizer}
\usage{
computeoptimizer(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
AWS Compute Optimizer is a service that analyzes the configuration and
utilization metrics of your AWS resources, such as EC2 instances and
Auto Scaling groups. It reports whether your resources are optimal, and
generates optimization recommendations to reduce the cost and improve
the performance of your workloads. Compute Optimizer also provides
recent utilization metric data, as well as projected utilization metric
data for the recommendations, which you can use to evaluate which
recommendation provides the best price-performance trade-off. The
analysis of your usage patterns can help you decide when to move or
resize your running resources, and still meet your performance and
capacity requirements. For more information about Compute Optimizer, see
the \href{https://docs.aws.amazon.com/compute-optimizer/latest/ug/}{AWS Compute Optimizer User Guide}.
}
\section{Service syntax}{
\preformatted{svc <- computeoptimizer(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=computeoptimizer_get_auto_scaling_group_recommendations]{get_auto_scaling_group_recommendations} \tab Returns Auto Scaling group recommendations \cr
\link[=computeoptimizer_get_ec2_instance_recommendations]{get_ec2_instance_recommendations} \tab Returns Amazon EC2 instance recommendations \cr
\link[=computeoptimizer_get_ec2_recommendation_projected_metrics]{get_ec2_recommendation_projected_metrics} \tab Returns the projected utilization metrics of Amazon EC2 instance recommendations \cr
\link[=computeoptimizer_get_enrollment_status]{get_enrollment_status} \tab Returns the enrollment (opt in) status of an account to the AWS Compute Optimizer service\cr
\link[=computeoptimizer_get_recommendation_summaries]{get_recommendation_summaries} \tab Returns the optimization findings for an account \cr
\link[=computeoptimizer_update_enrollment_status]{update_enrollment_status} \tab Updates the enrollment (opt in) status of an account to the AWS Compute Optimizer service
}
}

\examples{
\donttest{svc <- computeoptimizer()
svc$get_auto_scaling_group_recommendations(
  Foo = 123
)}

}
