% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_recommendation_summaries}
\alias{computeoptimizer_get_recommendation_summaries}
\title{Returns the optimization findings for an account}
\usage{
computeoptimizer_get_recommendation_summaries(accountIds, nextToken,
  maxResults)
}
\arguments{
\item{accountIds}{The IDs of the AWS accounts for which to return recommendation
summaries.

If your account is the management account of an organization, use this
parameter to specify the member accounts for which you want to return
recommendation summaries.

Only one account ID can be specified per request.}

\item{nextToken}{The token to advance to the next page of recommendation summaries.}

\item{maxResults}{The maximum number of recommendation summaries to return with a single
request.

To retrieve the remaining results, make another request with the
returned \code{NextToken} value.}
}
\description{
Returns the optimization findings for an account.

For example, it returns the number of Amazon EC2 instances in an account
that are under-provisioned, over-provisioned, or optimized. It also
returns the number of Auto Scaling groups in an account that are not
optimized, or optimized.
}
\section{Request syntax}{
\preformatted{svc$get_recommendation_summaries(
  accountIds = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
