% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_organization_config_rule_detailed_status}
\alias{configservice_get_organization_config_rule_detailed_status}
\title{Get organization config rule detailed status}
\usage{
configservice_get_organization_config_rule_detailed_status(
  OrganizationConfigRuleName, Filters, Limit, NextToken)
}
\arguments{
\item{OrganizationConfigRuleName}{[required]}

\item{Filters}{}

\item{Limit}{}

\item{NextToken}{}
}
\description{
Get organization config rule detailed status
}
\section{Request syntax}{
\preformatted{svc$get_organization_config_rule_detailed_status(
  OrganizationConfigRuleName = "string",
  Filters = list(
    AccountId = "string",
    MemberAccountRuleStatus = "CREATE_SUCCESSFUL"|"CREATE_IN_PROGRESS"|"CREATE_FAILED"|"UPDATE_SUCCESSFUL"|"UPDATE_FAILED"|"UPDATE_IN_PROGRESS"|"DELETE_SUCCESSFUL"|"DELETE_FAILED"|"DELETE_IN_PROGRESS"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
