% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcontactlens_service.R
\name{connectcontactlens}
\alias{connectcontactlens}
\title{Amazon Connect Contact Lens}
\usage{
connectcontactlens(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Contact Lens for Amazon Connect enables you to analyze conversations
between customer and agents, by using speech transcription, natural
language processing, and intelligent search capabilities. It performs
sentiment analysis, detects issues, and enables you to automatically
categorize contacts.

Contact Lens for Amazon Connect provides both real-time and post-call
analytics of customer-agent conversations. For more information, see
\href{https://docs.aws.amazon.com/connect/latest/adminguide/analyze-conversations.html}{Analyze conversations using Contact Lens}
in the \emph{Amazon Connect Administrator Guide}.
}
\section{Service syntax}{
\preformatted{svc <- connectcontactlens(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=connectcontactlens_list_realtime_contact_analysis_segments]{list_realtime_contact_analysis_segments} \tab Provides a list of analysis segments for a real-time analysis session
}
}

\examples{
\dontrun{
svc <- connectcontactlens()
svc$list_realtime_contact_analysis_segments(
  Foo = 123
)
}

}
