% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_service.R
\name{connectparticipant}
\alias{connectparticipant}
\title{Amazon Connect Participant Service}
\usage{
connectparticipant(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Connect is a cloud-based contact center solution that makes it
easy to set up and manage a customer contact center and provide reliable
customer engagement at any scale.

Amazon Connect enables customer contacts through voice or chat.

The APIs described here are used by chat participants, such as agents
and customers.
}
\section{Service syntax}{
\preformatted{svc <- connectparticipant(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=connectparticipant_create_participant_connection]{create_participant_connection} \tab Creates the participant's connection \cr
\link[=connectparticipant_disconnect_participant]{disconnect_participant} \tab Disconnects a participant \cr
\link[=connectparticipant_get_transcript]{get_transcript} \tab Retrieves a transcript of the session\cr
\link[=connectparticipant_send_event]{send_event} \tab Sends an event \cr
\link[=connectparticipant_send_message]{send_message} \tab Sends a message
}
}

\examples{
\dontrun{svc <- connectparticipant()
svc$create_participant_connection(
  Foo = 123
)}

}
