% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{controltower}
\alias{controltower}
\title{AWS Control Tower}
\usage{
controltower(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}:} {\itemize{
\item{\strong{creds}:} {\itemize{
\item{\strong{access_key_id}:} {AWS access key ID}
\item{\strong{secret_access_key}:} {AWS secret access key}
\item{\strong{session_token}:} {AWS temporary session token}
}}
\item{\strong{profile}:} {The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}:} {Set anonymous credentials.}
}}
\item{\strong{endpoint}:} {The complete URL to use for the constructed client.}
\item{\strong{region}:} {The AWS Region used in instantiating the client.}
\item{\strong{close_connection}:} {Immediately close all HTTP connections.}
\item{\strong{timeout}:} {The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}:} {Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}:} {Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}:} {\itemize{
\item{\strong{access_key_id}:} {AWS access key ID}
\item{\strong{secret_access_key}:} {AWS secret access key}
\item{\strong{session_token}:} {AWS temporary session token}
}}
\item{\strong{profile}:} {The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}:} {Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
These interfaces allow you to apply the AWS library of pre-defined
\emph{controls} to your organizational units, programmatically. In AWS
Control Tower, the terms "control" and "guardrail" are synonyms.

To call these APIs, you'll need to know:
\itemize{
\item the \code{controlIdentifier} for the control--or guardrail--you are
targeting.
\item the ARN associated with the target organizational unit (OU), which
we call the \code{targetIdentifier}.
\item the ARN associated with a resource that you wish to tag or untag.
}

\strong{To get the \code{controlIdentifier} for your AWS Control Tower control:}

The \code{controlIdentifier} is an ARN that is specified for each control.
You can view the \code{controlIdentifier} in the console on the \strong{Control
details} page, as well as in the documentation.

The \code{controlIdentifier} is unique in each AWS Region for each control.
You can find the \code{controlIdentifier} for each Region and control in the
\href{https://docs.aws.amazon.com/controltower/latest/userguide/control-metadata-tables.html}{Tables of control metadata}
in the \emph{AWS Control Tower User Guide.}

A quick-reference list of control identifers for the AWS Control Tower
legacy \emph{Strongly recommended} and \emph{Elective} controls is given in
\href{https://docs.aws.amazon.com/controltower/latest/userguide/}{Resource identifiers for APIs and controls} in
the \href{https://docs.aws.amazon.com/controltower/latest/userguide/control-identifiers.html}{Controls reference guide section}
of the \emph{AWS Control Tower User Guide}. Remember that \emph{Mandatory}
controls cannot be added or removed.

\strong{ARN format:} \verb{arn:aws:controltower:\{REGION\}::control/\{CONTROL_NAME\}}

\strong{Example:}

\verb{arn:aws:controltower:us-west-2::control/AWS-GR_AUTOSCALING_LAUNCH_CONFIG_PUBLIC_IP_DISABLED}

\strong{To get the \code{targetIdentifier}:}

The \code{targetIdentifier} is the ARN for an OU.

In the AWS Organizations console, you can find the ARN for the OU on the
\strong{Organizational unit details} page associated with that OU.

\strong{OU ARN format:}

\verb{arn:$\{Partition\}:organizations::$\{MasterAccountId\}:ou/o-$\{OrganizationId\}/ou-$\{OrganizationalUnitId\}}

\strong{Details and examples}
\itemize{
\item \href{https://docs.aws.amazon.com/controltower/latest/userguide/control-api-examples-short.html}{Control API input and output examples with CLI}
\item \href{https://docs.aws.amazon.com/controltower/latest/userguide/enable-controls.html}{Enable controls with CloudFormation}
\item \href{https://docs.aws.amazon.com/controltower/latest/userguide/control-metadata-tables.html}{Control metadata tables}
\item \href{https://docs.aws.amazon.com/controltower/latest/userguide/control-identifiers.html}{List of identifiers for legacy controls}
\item \href{https://docs.aws.amazon.com/controltower/latest/userguide/controls.html}{Controls reference guide}
\item \href{https://docs.aws.amazon.com/controltower/latest/userguide/controls-reference.html}{Controls library groupings}
\item \href{https://docs.aws.amazon.com/controltower/latest/userguide/creating-resources-with-cloudformation.html}{Creating AWS Control Tower resources with AWS CloudFormation}
}

To view the open source resource repository on GitHub, see
\href{https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-controltower}{aws-cloudformation/aws-cloudformation-resource-providers-controltower}

\strong{Recording API Requests}

AWS Control Tower supports AWS CloudTrail, a service that records AWS
API calls for your AWS account and delivers log files to an Amazon S3
bucket. By using information collected by CloudTrail, you can determine
which requests the AWS Control Tower service received, who made the
request and when, and so on. For more about AWS Control Tower and its
support for CloudTrail, see \href{https://docs.aws.amazon.com/controltower/latest/userguide/logging-using-cloudtrail.html}{Logging AWS Control Tower Actions with AWS CloudTrail}
in the AWS Control Tower User Guide. To learn more about CloudTrail,
including how to turn it on and find your log files, see the AWS
CloudTrail User Guide.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- controltower(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.management:controltower_disable_control]{disable_control} \tab This API call turns off a control\cr
\link[paws.management:controltower_enable_control]{enable_control} \tab This API call activates a control\cr
\link[paws.management:controltower_get_control_operation]{get_control_operation} \tab Returns the status of a particular EnableControl or DisableControl operation\cr
\link[paws.management:controltower_get_enabled_control]{get_enabled_control} \tab Retrieves details about an enabled control\cr
\link[paws.management:controltower_list_enabled_controls]{list_enabled_controls} \tab Lists the controls enabled by AWS Control Tower on the specified organizational unit and the accounts it contains\cr
\link[paws.management:controltower_list_tags_for_resource]{list_tags_for_resource} \tab Returns a list of tags associated with the resource\cr
\link[paws.management:controltower_tag_resource]{tag_resource} \tab Applies tags to a resource\cr
\link[paws.management:controltower_untag_resource]{untag_resource} \tab Removes tags from a resource
}
}

\examples{
\dontrun{
svc <- controltower()
svc$disable_control(
  Foo = 123
)
}

}
