% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_service.R
\name{costandusagereportservice}
\alias{costandusagereportservice}
\title{AWS Cost and Usage Report Service}
\usage{
costandusagereportservice()
}
\description{
All public APIs for AWS Cost and Usage Report service
}
\section{Example}{
\preformatted{costandusagereportservice <- paws::costandusagereportservice()
costandusagereportservice$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=costandusagereportservice_delete_report_definition]{delete_report_definition} \tab Delete a specified report definition \cr
\link[=costandusagereportservice_describe_report_definitions]{describe_report_definitions} \tab Describe a list of report definitions owned by the account\cr
\link[=costandusagereportservice_put_report_definition]{put_report_definition} \tab Create a new report definition
}
}

