% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_delete_report_definition}
\alias{costandusagereportservice_delete_report_definition}
\title{Delete a specified report definition}
\usage{
costandusagereportservice_delete_report_definition(ReportName)
}
\arguments{
\item{ReportName}{}
}
\description{
Delete a specified report definition
}
\section{Request syntax}{
\preformatted{costandusagereportservice$delete_report_definition(
  ReportName = "string"
)
}
}

\keyword{internal}
