% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_describe_report_definitions}
\alias{costandusagereportservice_describe_report_definitions}
\title{Describe a list of report definitions owned by the account}
\usage{
costandusagereportservice_describe_report_definitions(MaxResults,
  NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Describe a list of report definitions owned by the account
}
\section{Request syntax}{
\preformatted{costandusagereportservice$describe_report_definitions(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
