% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_add_tags_to_resource}
\alias{databasemigrationservice_add_tags_to_resource}
\title{Adds metadata tags to an AWS DMS resource, including replication
instance, endpoint, security group, and migration task}
\usage{
databasemigrationservice_add_tags_to_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] Identifies the AWS DMS resource to which tags should be added. The value
for this parameter is an Amazon Resource Name (ARN).

For AWS DMS, you can tag a replication instance, an endpoint, or a
replication task.}

\item{Tags}{[required] One or more tags to be assigned to the resource.}
}
\description{
Adds metadata tags to an AWS DMS resource, including replication
instance, endpoint, security group, and migration task. These tags can
also be used with cost allocation reporting to track cost associated
with DMS resources, or used in a Condition statement in an IAM policy
for DMS.
}
\section{Request syntax}{
\preformatted{svc$add_tags_to_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\examples{
# Adds metadata tags to an AWS DMS resource, including replication
# instance, endpoint, security group, and migration task. These tags can
# also be used with cost allocation reporting to track cost associated
# with AWS DMS resources, or used in a Condition statement in an IAM
# policy for AWS DMS.
\donttest{svc$add_tags_to_resource(
  ResourceArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7E",
  Tags = list(
    list(
      Key = "Acount",
      Value = "1633456"
    )
  )
)}

}
\keyword{internal}
