% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_modify_replication_instance}
\alias{databasemigrationservice_modify_replication_instance}
\title{Modifies the replication instance to apply new settings}
\usage{
databasemigrationservice_modify_replication_instance(
  ReplicationInstanceArn, AllocatedStorage, ApplyImmediately,
  ReplicationInstanceClass, VpcSecurityGroupIds,
  PreferredMaintenanceWindow, MultiAZ, EngineVersion,
  AllowMajorVersionUpgrade, AutoMinorVersionUpgrade,
  ReplicationInstanceIdentifier)
}
\arguments{
\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance.}

\item{AllocatedStorage}{The amount of storage (in gigabytes) to be allocated for the replication
instance.}

\item{ApplyImmediately}{Indicates whether the changes should be applied immediately or during
the next maintenance window.}

\item{ReplicationInstanceClass}{The compute and memory capacity of the replication instance as defined
for the specified replication instance class. For example to specify the
instance class dms.c4.large, set this parameter to \code{"dms.c4.large"}.

For more information on the settings and capacities for the available
replication instance classes, see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_ReplicationInstance.html#CHAP_ReplicationInstance.InDepth}{Selecting the right AWS DMS replication instance for your migration}.}

\item{VpcSecurityGroupIds}{Specifies the VPC security group to be used with the replication
instance. The VPC security group must work with the VPC containing the
replication instance.}

\item{PreferredMaintenanceWindow}{The weekly time range (in UTC) during which system maintenance can
occur, which might result in an outage. Changing this parameter does not
result in an outage, except in the following situation, and the change
is asynchronously applied as soon as possible. If moving this window to
the current time, there must be at least 30 minutes between the current
time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes}

\item{MultiAZ}{Specifies whether the replication instance is a Multi-AZ deployment. You
can't set the \code{AvailabilityZone} parameter if the Multi-AZ parameter is
set to \code{true}.}

\item{EngineVersion}{The engine version number of the replication instance.

When modifying a major engine version of an instance, also set
\code{AllowMajorVersionUpgrade} to \code{true}.}

\item{AllowMajorVersionUpgrade}{Indicates that major version upgrades are allowed. Changing this
parameter does not result in an outage, and the change is asynchronously
applied as soon as possible.

This parameter must be set to \code{true} when specifying a value for the
\code{EngineVersion} parameter that is a different major version than the
replication instance's current version.}

\item{AutoMinorVersionUpgrade}{A value that indicates that minor version upgrades are applied
automatically to the replication instance during the maintenance window.
Changing this parameter doesn't result in an outage, except in the case
described following. The change is asynchronously applied as soon as
possible.

An outage does result if these factors apply:
\itemize{
\item This parameter is set to \code{true} during the maintenance window.
\item A newer minor version is available.
\item AWS DMS has enabled automatic patching for the given engine version.
}}

\item{ReplicationInstanceIdentifier}{The replication instance identifier. This parameter is stored as a
lowercase string.}
}
\description{
Modifies the replication instance to apply new settings. You can change
one or more parameters by specifying these parameters and the new values
in the request.
}
\details{
Some settings are applied during the maintenance window.
}
\section{Request syntax}{
\preformatted{svc$modify_replication_instance(
  ReplicationInstanceArn = "string",
  AllocatedStorage = 123,
  ApplyImmediately = TRUE|FALSE,
  ReplicationInstanceClass = "string",
  VpcSecurityGroupIds = list(
    "string"
  ),
  PreferredMaintenanceWindow = "string",
  MultiAZ = TRUE|FALSE,
  EngineVersion = "string",
  AllowMajorVersionUpgrade = TRUE|FALSE,
  AutoMinorVersionUpgrade = TRUE|FALSE,
  ReplicationInstanceIdentifier = "string"
)
}
}

\examples{
\dontrun{
# Modifies the replication instance to apply new settings. You can change
# one or more parameters by specifying these parameters and the new values
# in the request. Some settings are applied during the maintenance window.
svc$modify_replication_instance(
  AllocatedStorage = 123L,
  AllowMajorVersionUpgrade = TRUE,
  ApplyImmediately = TRUE,
  AutoMinorVersionUpgrade = TRUE,
  EngineVersion = "1.5.0",
  MultiAZ = TRUE,
  PreferredMaintenanceWindow = "sun:06:00-sun:14:00",
  ReplicationInstanceArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJ...",
  ReplicationInstanceClass = "dms.t2.micro",
  ReplicationInstanceIdentifier = "test-rep-1",
  VpcSecurityGroupIds = list()
)
}

}
\keyword{internal}
