% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_remove_tags_from_resource}
\alias{databasemigrationservice_remove_tags_from_resource}
\title{Removes metadata tags from a DMS resource}
\usage{
databasemigrationservice_remove_tags_from_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] \>The Amazon Resource Name (ARN) of the AWS DMS resource the tag is to
be removed from.}

\item{TagKeys}{[required] The tag key (name) of the tag to be removed.}
}
\description{
Removes metadata tags from a DMS resource.
}
\section{Request syntax}{
\preformatted{databasemigrationservice$remove_tags_from_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\examples{
# Removes metadata tags from an AWS DMS resource.
\donttest{databasemigrationservice$remove_tags_from_resource(
  ResourceArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7E",
  TagKeys = list()
)}

}
\keyword{internal}
