% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_start_task_execution}
\alias{datasync_start_task_execution}
\title{Starts a specific invocation of a task}
\usage{
datasync_start_task_execution(TaskArn, OverrideOptions)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the task to start.}

\item{OverrideOptions}{}
}
\description{
Starts a specific invocation of a task. A \code{TaskExecution} value
represents an individual run of a task. Each task can have at most one
\code{TaskExecution} at a time.
}
\details{
\code{TaskExecution} has the following transition phases: INITIALIZING \|
PREPARING \| TRANSFERRING \| VERIFYING \| SUCCESS/FAILURE.

For detailed information, see \emph{Task Execution} in Components and
Terminology in the \emph{AWS DataSync User Guide}.
}
\section{Request syntax}{
\preformatted{svc$start_task_execution(
  TaskArn = "string",
  OverrideOptions = list(
    VerifyMode = "POINT_IN_TIME_CONSISTENT"|"NONE",
    Atime = "NONE"|"BEST_EFFORT",
    Mtime = "NONE"|"PRESERVE",
    Uid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    Gid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    PreserveDeletedFiles = "PRESERVE"|"REMOVE",
    PreserveDevices = "NONE"|"PRESERVE",
    PosixPermissions = "NONE"|"BEST_EFFORT"|"PRESERVE",
    BytesPerSecond = 123
  )
)
}
}

\keyword{internal}
