% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_device}
\alias{devicefarm_get_device}
\title{Gets information about a unique device type}
\usage{
devicefarm_get_device(arn)
}
\arguments{
\item{arn}{[required] The device type\'s ARN.}
}
\description{
Gets information about a unique device type.
}
\section{Request syntax}{
\preformatted{svc$get_device(
  arn = "string"
)
}
}

\section{Examples}{
\preformatted{# The following example returns information about a specific device.
svc$get_device(
  arn = "arn:aws:devicefarm:us-west-2::device:123EXAMPLE"
)
}
}

\keyword{internal}
