% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_service.R
\name{devopsguru}
\alias{devopsguru}
\title{Amazon DevOps Guru}
\usage{
devopsguru(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon DevOps Guru is a fully managed service that helps you identify
anomalous behavior in business critical operational applications. You
specify the AWS resources that you want DevOps Guru to cover, then the
Amazon CloudWatch metrics and AWS CloudTrail events related to those
resources are analyzed. When anomalous behavior is detected, DevOps Guru
creates an \emph{insight} that includes recommendations, related events, and
related metrics that can help you improve your operational applications.
For more information, see \href{https://docs.aws.amazon.com/devops-guru/latest/userguide/welcome.html}{What is Amazon DevOps Guru}.

You can specify 1 or 2 Amazon Simple Notification Service topics so you
are notified every time a new insight is created. You can also enable
DevOps Guru to generate an OpsItem in AWS Systems Manager for each
insight to help you manage and track your work addressing insights.

To learn about the DevOps Guru workflow, see \href{https://docs.aws.amazon.com/devops-guru/latest/userguide/welcome.html#how-it-works}{How DevOps Guru works}.
To learn about DevOps Guru concepts, see \href{https://docs.aws.amazon.com/devops-guru/latest/userguide/concepts.html}{Concepts in DevOps Guru}.
}
\section{Service syntax}{
\preformatted{svc <- devopsguru(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=devopsguru_add_notification_channel]{add_notification_channel} \tab Adds a notification channel to DevOps Guru\cr
\link[=devopsguru_describe_account_health]{describe_account_health} \tab Returns the number of open reactive insights, the number of open proactive insights, and the number of metrics analyzed in your AWS account\cr
\link[=devopsguru_describe_account_overview]{describe_account_overview} \tab For the time range passed in, returns the number of open reactive insight that were created, the number of open proactive insights that were created, and the Mean Time to Recover (MTTR) for all closed reactive insights\cr
\link[=devopsguru_describe_anomaly]{describe_anomaly} \tab Returns details about an anomaly that you specify using its ID\cr
\link[=devopsguru_describe_insight]{describe_insight} \tab Returns details about an insight that you specify using its ID\cr
\link[=devopsguru_describe_resource_collection_health]{describe_resource_collection_health} \tab Returns the number of open proactive insights, open reactive insights, and the Mean Time to Recover (MTTR) for all closed insights in resource collections in your account\cr
\link[=devopsguru_describe_service_integration]{describe_service_integration} \tab Returns the integration status of services that are integrated with DevOps Guru\cr
\link[=devopsguru_get_resource_collection]{get_resource_collection} \tab Returns lists AWS resources that are of the specified resource collection type\cr
\link[=devopsguru_list_anomalies_for_insight]{list_anomalies_for_insight} \tab Returns a list of the anomalies that belong to an insight that you specify using its ID\cr
\link[=devopsguru_list_events]{list_events} \tab Returns a list of the events emitted by the resources that are evaluated by DevOps Guru\cr
\link[=devopsguru_list_insights]{list_insights} \tab Returns a list of insights in your AWS account\cr
\link[=devopsguru_list_notification_channels]{list_notification_channels} \tab Returns a list of notification channels configured for DevOps Guru\cr
\link[=devopsguru_list_recommendations]{list_recommendations} \tab Returns a list of a specified insight's recommendations\cr
\link[=devopsguru_put_feedback]{put_feedback} \tab Collects customer feedback about the specified insight\cr
\link[=devopsguru_remove_notification_channel]{remove_notification_channel} \tab Removes a notification channel from DevOps Guru\cr
\link[=devopsguru_search_insights]{search_insights} \tab Returns a list of insights in your AWS account\cr
\link[=devopsguru_update_resource_collection]{update_resource_collection} \tab Updates the collection of resources that DevOps Guru analyzes\cr
\link[=devopsguru_update_service_integration]{update_service_integration} \tab Enables or disables integration with a service that can be integrated with DevOps Guru
}
}

\examples{
\dontrun{
svc <- devopsguru()
svc$add_notification_channel(
  Foo = 123
)
}

}
