% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_direct_connect_gateway_association}
\alias{directconnect_delete_direct_connect_gateway_association}
\title{Deletes the association between the specified Direct Connect gateway and
virtual private gateway}
\usage{
directconnect_delete_direct_connect_gateway_association(
  directConnectGatewayId, virtualGatewayId)
}
\arguments{
\item{directConnectGatewayId}{[required] The ID of the Direct Connect gateway.}

\item{virtualGatewayId}{[required] The ID of the virtual private gateway.}
}
\description{
Deletes the association between the specified Direct Connect gateway and
virtual private gateway.
}
\section{Request syntax}{
\preformatted{directconnect$delete_direct_connect_gateway_association(
  directConnectGatewayId = "string",
  virtualGatewayId = "string"
)
}
}

\keyword{internal}
