% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_global_table_settings}
\alias{dynamodb_describe_global_table_settings}
\title{Describes Region-specific settings for a global table}
\usage{
dynamodb_describe_global_table_settings(GlobalTableName)
}
\arguments{
\item{GlobalTableName}{[required] The name of the global table to describe.}
}
\description{
Describes Region-specific settings for a global table.
}
\details{
This method only applies to \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/globaltables.V1.html}{Version 2017.11.29}
of global tables.
}
\section{Request syntax}{
\preformatted{svc$describe_global_table_settings(
  GlobalTableName = "string"
)
}
}

\keyword{internal}
