% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebs_service.R
\name{ebs}
\alias{ebs}
\title{Amazon Elastic Block Store}
\usage{
ebs(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
You can use the Amazon Elastic Block Store (EBS) direct APIs to directly
read the data on your EBS snapshots, and identify the difference between
two snapshots. You can view the details of blocks in an EBS snapshot,
compare the block difference between two snapshots, and directly access
the data in a snapshot. If you're an independent software vendor (ISV)
who offers backup services for EBS, the EBS direct APIs make it easier
and more cost-effective to track incremental changes on your EBS volumes
via EBS snapshots. This can be done without having to create new volumes
from EBS snapshots.

This API reference provides detailed information about the actions, data
types, parameters, and errors of the EBS direct APIs. For more
information about the elements that make up the EBS direct APIs, and
examples of how to use them effectively, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ebs-accessing-snapshot.html}{Accessing the Contents of an EBS Snapshot}
in the \emph{Amazon Elastic Compute Cloud User Guide}. For more information
about the supported AWS Regions, endpoints, and service quotas for the
EBS direct APIs, see \href{https://docs.aws.amazon.com/general/latest/gr/ebs-service.html}{Amazon Elastic Block Store Endpoints and Quotas}
in the \emph{AWS General Reference}.
}
\section{Service syntax}{
\preformatted{svc <- ebs(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=ebs_complete_snapshot]{complete_snapshot} \tab Seals and completes the snapshot after all of the required blocks of data have been written to it \cr
\link[=ebs_get_snapshot_block]{get_snapshot_block} \tab Returns the data in a block in an Amazon Elastic Block Store snapshot \cr
\link[=ebs_list_changed_blocks]{list_changed_blocks} \tab Returns the block indexes and block tokens for blocks that are different between two Amazon Elastic Block Store snapshots of the same volume/snapshot lineage\cr
\link[=ebs_list_snapshot_blocks]{list_snapshot_blocks} \tab Returns the block indexes and block tokens for blocks in an Amazon Elastic Block Store snapshot \cr
\link[=ebs_put_snapshot_block]{put_snapshot_block} \tab Writes a block of data to a block in the snapshot \cr
\link[=ebs_start_snapshot]{start_snapshot} \tab Creates a new Amazon EBS snapshot
}
}

\examples{
\dontrun{
svc <- ebs()
svc$complete_snapshot(
  Foo = 123
)
}

}
