% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_put_account_setting_default}
\alias{ecs_put_account_setting_default}
\title{Modifies an account setting for all IAM users on an account for whom no
individual account setting has been specified}
\usage{
ecs_put_account_setting_default(name, value)
}
\arguments{
\item{name}{[required] The resource name for which to modify the account setting. If
\code{serviceLongArnFormat} is specified, the ARN for your Amazon ECS
services is affected. If \code{taskLongArnFormat} is specified, the ARN and
resource ID for your Amazon ECS tasks is affected. If
\code{containerInstanceLongArnFormat} is specified, the ARN and resource ID
for your Amazon ECS container instances is affected. If \code{awsvpcTrunking}
is specified, the ENI limit for your Amazon ECS container instances is
affected. If \code{containerInsights} is specified, the default setting for
CloudWatch Container Insights for your clusters is affected.}

\item{value}{[required] The account setting value for the specified principal ARN. Accepted
values are \code{enabled} and \code{disabled}.}
}
\description{
Modifies an account setting for all IAM users on an account for whom no
individual account setting has been specified.
}
\section{Request syntax}{
\preformatted{svc$put_account_setting_default(
  name = "serviceLongArnFormat"|"taskLongArnFormat"|"containerInstanceLongArnFormat"|"awsvpcTrunking"|"containerInsights",
  value = "string"
)
}
}

\examples{
# This example modifies the default account setting for the specified
# resource for all IAM users or roles on an account. These changes apply
# to the entire AWS account, unless an IAM user or role explicitly
# overrides these settings for themselves.
\donttest{svc$put_account_setting_default(
  name = "serviceLongArnFormat",
  value = "enabled"
)}

}
\keyword{internal}
