% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_rules}
\alias{eventbridge_list_rules}
\title{Lists your Amazon EventBridge rules}
\usage{
eventbridge_list_rules(NamePrefix, EventBusName, NextToken, Limit)
}
\arguments{
\item{NamePrefix}{The prefix matching the rule name.}

\item{EventBusName}{The name or ARN of the event bus to list the rules for. If you omit
this, the default event bus is used.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{Limit}{The maximum number of results to return.}
}
\description{
Lists your Amazon EventBridge rules. You can either list all the rules
or you can provide a prefix to match to the rule names.

ListRules does not list the targets of a rule. To see the targets
associated with a rule, use ListTargetsByRule.
}
\section{Request syntax}{
\preformatted{svc$list_rules(
  NamePrefix = "string",
  EventBusName = "string",
  NextToken = "string",
  Limit = 123
)
}
}

\keyword{internal}
