% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_put_policy}
\alias{fms_put_policy}
\title{Creates an AWS Firewall Manager policy}
\usage{
fms_put_policy(Policy)
}
\arguments{
\item{Policy}{[required] The details of the AWS Firewall Manager policy to be created.}
}
\description{
Creates an AWS Firewall Manager policy.
}
\section{Request syntax}{
\preformatted{fms$put_policy(
  Policy = list(
    PolicyId = "string",
    PolicyName = "string",
    PolicyUpdateToken = "string",
    SecurityServicePolicyData = list(
      Type = "WAF",
      ManagedServiceData = "string"
    ),
    ResourceType = "string",
    ResourceTags = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    ExcludeResourceTags = TRUE|FALSE,
    RemediationEnabled = TRUE|FALSE,
    IncludeMap = list(
      list(
        "string"
      )
    ),
    ExcludeMap = list(
      list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
