% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastqueryservice_service.R
\name{forecastqueryservice}
\alias{forecastqueryservice}
\title{Amazon Forecast Query Service}
\usage{
forecastqueryservice(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Provides APIs for creating and managing Amazon Forecast resources.
}
\section{Service syntax}{
\preformatted{svc <- forecastqueryservice(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Examples}{
\preformatted{svc$query_forecast(
  Foo = 123
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=forecastqueryservice_query_forecast]{query_forecast} \tab Retrieves a forecast filtered by the supplied criteria
}
}

