% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_create_forecast_export_job}
\alias{forecastservice_create_forecast_export_job}
\title{Exports a forecast created by the CreateForecast operation to your
Amazon Simple Storage Service (Amazon S3) bucket}
\usage{
forecastservice_create_forecast_export_job(ForecastExportJobName,
  ForecastArn, Destination)
}
\arguments{
\item{ForecastExportJobName}{[required] The name for the forecast export job.}

\item{ForecastArn}{[required] The Amazon Resource Name (ARN) of the forecast that you want to export.}

\item{Destination}{[required] The location where you want to save the forecast and an AWS Identity and
Access Management (IAM) role that Amazon Forecast can assume to access
the location. The forecast must be exported to an Amazon S3 bucket.

If encryption is used, \code{Destination} must include an AWS Key Management
Service (KMS) key. The IAM role must allow Amazon Forecast permission to
access the key.}
}
\description{
Exports a forecast created by the CreateForecast operation to your
Amazon Simple Storage Service (Amazon S3) bucket. The forecast file name
will match the following conventions:
}
\details{
\\<ForecastExportJobName\\>\\\emph{\\<ExportTimestamp\\>\\}\\<PageNumber\\>

where the \\<ExportTimestamp\\> component is in Java SimpleDateFormat
(yyyy-MM-ddTHH-mm-ssZ).

You must specify a DataDestination object that includes an AWS Identity
and Access Management (IAM) role that Amazon Forecast can assume to
access the Amazon S3 bucket. For more information, see
aws-forecast-iam-roles.

For more information, see howitworks-forecast.

To get a list of all your forecast export jobs, use the
ListForecastExportJobs operation.

The \code{Status} of the forecast export job must be \code{ACTIVE} before you can
access the forecast in your Amazon S3 bucket. To get the status, use the
DescribeForecastExportJob operation.
}
\section{Request syntax}{
\preformatted{svc$create_forecast_export_job(
  ForecastExportJobName = "string",
  ForecastArn = "string",
  Destination = list(
    S3Config = list(
      Path = "string",
      RoleArn = "string",
      KMSKeyArn = "string"
    )
  )
)
}
}

\keyword{internal}
