% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_create_detector_version}
\alias{frauddetector_create_detector_version}
\title{Creates a detector version}
\usage{
frauddetector_create_detector_version(detectorId, description,
  externalModelEndpoints, rules, modelVersions)
}
\arguments{
\item{detectorId}{[required] The ID of the detector under which you want to create a new version.}

\item{description}{The description of the detector version.}

\item{externalModelEndpoints}{The Amazon Sagemaker model endpoints to include in the detector version.}

\item{rules}{[required] The rules to include in the detector version.}

\item{modelVersions}{The model versions to include in the detector version.}
}
\description{
Creates a detector version. The detector version starts in a \code{DRAFT}
status.
}
\section{Request syntax}{
\preformatted{svc$create_detector_version(
  detectorId = "string",
  description = "string",
  externalModelEndpoints = list(
    "string"
  ),
  rules = list(
    list(
      detectorId = "string",
      ruleId = "string",
      ruleVersion = "string"
    )
  ),
  modelVersions = list(
    list(
      modelId = "string",
      modelType = "ONLINE_FRAUD_INSIGHTS",
      modelVersionNumber = "string"
    )
  )
)
}
}

\keyword{internal}
