% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_create_model_version}
\alias{frauddetector_create_model_version}
\title{Creates a version of the model using the specified model type}
\usage{
frauddetector_create_model_version(modelId, modelType, description)
}
\arguments{
\item{modelId}{[required] The model ID.}

\item{modelType}{[required] The model type.}

\item{description}{The model version description.}
}
\description{
Creates a version of the model using the specified model type.
}
\section{Request syntax}{
\preformatted{svc$create_model_version(
  modelId = "string",
  modelType = "ONLINE_FRAUD_INSIGHTS",
  description = "string"
)
}
}

\keyword{internal}
