% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_detector}
\alias{frauddetector_delete_detector}
\title{Deletes the detector}
\usage{
frauddetector_delete_detector(detectorId)
}
\arguments{
\item{detectorId}{[required] The ID of the detector to delete.}
}
\description{
Deletes the detector. Before deleting a detector, you must first delete
all detector versions and rule versions associated with the detector.
}
\details{
When you delete a detector, Amazon Fraud Detector permanently deletes
the detector and the data is no longer stored in Amazon Fraud Detector.
}
\section{Request syntax}{
\preformatted{svc$delete_detector(
  detectorId = "string"
)
}
}

\keyword{internal}
