% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_models}
\alias{frauddetector_get_models}
\title{Gets all of the models for the AWS account, or the specified model type,
or gets a single model for the specified model type, model ID
combination}
\usage{
frauddetector_get_models(modelType, modelId, nextToken, maxResults)
}
\arguments{
\item{modelType}{The model type.}

\item{modelId}{The model ID.}

\item{nextToken}{The next token for the request.}

\item{maxResults}{The maximum results to return for the request.}
}
\description{
Gets all of the models for the AWS account, or the specified model type,
or gets a single model for the specified model type, model ID
combination.
}
\section{Request syntax}{
\preformatted{svc$get_models(
  modelType = "ONLINE_FRAUD_INSIGHTS",
  modelId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
