% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_rules}
\alias{frauddetector_get_rules}
\title{Get all rules for a detector (paginated) if ruleId and ruleVersion are
not specified}
\usage{
frauddetector_get_rules(ruleId, detectorId, ruleVersion, nextToken,
  maxResults)
}
\arguments{
\item{ruleId}{The rule ID.}

\item{detectorId}{[required] The detector ID.}

\item{ruleVersion}{The rule version.}

\item{nextToken}{The next page token.}

\item{maxResults}{The maximum number of rules to return for the request.}
}
\description{
Get all rules for a detector (paginated) if \code{ruleId} and \code{ruleVersion}
are not specified. Gets all rules for the detector and the \code{ruleId} if
present (paginated). Gets a specific rule if both the \code{ruleId} and the
\code{ruleVersion} are specified.

This is a paginated API. Providing null maxResults results in retrieving
maximum of 100 records per page. If you provide maxResults the value
must be between 50 and 100. To get the next page result, a provide a
pagination token from GetRulesResult as part of your request. Null
pagination token fetches the records from the beginning.
}
\section{Request syntax}{
\preformatted{svc$get_rules(
  ruleId = "string",
  detectorId = "string",
  ruleVersion = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
