% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_model_version}
\alias{frauddetector_update_model_version}
\title{Updates a model version}
\usage{
frauddetector_update_model_version(modelId, modelType,
  majorVersionNumber, externalEventsDetail, tags)
}
\arguments{
\item{modelId}{[required] The model ID.}

\item{modelType}{[required] The model type.}

\item{majorVersionNumber}{[required] The major version number.}

\item{externalEventsDetail}{The event details.}

\item{tags}{A collection of key and value pairs.}
}
\description{
Updates a model version. Updating a model version retrains an existing
model version using updated training data and produces a new minor
version of the model. You can update the training data set location and
data access role attributes using this action. This action creates and
trains a new minor version of the model, for example version 1.01, 1.02,
1.03.
}
\section{Request syntax}{
\preformatted{svc$update_model_version(
  modelId = "string",
  modelType = "ONLINE_FRAUD_INSIGHTS",
  majorVersionNumber = "string",
  externalEventsDetail = list(
    dataLocation = "string",
    dataAccessRoleArn = "string"
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
