% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_rule_version}
\alias{frauddetector_update_rule_version}
\title{Updates a rule version resulting in a new rule version}
\usage{
frauddetector_update_rule_version(rule, description, expression,
  language, outcomes, tags)
}
\arguments{
\item{rule}{[required] The rule to update.}

\item{description}{The description.}

\item{expression}{[required] The rule expression.}

\item{language}{[required] The language.}

\item{outcomes}{[required] The outcomes.}

\item{tags}{The tags to assign to the rule version.}
}
\description{
Updates a rule version resulting in a new rule version. Updates a rule
version resulting in a new rule version (version 1, 2, 3 ...).
}
\section{Request syntax}{
\preformatted{svc$update_rule_version(
  rule = list(
    detectorId = "string",
    ruleId = "string",
    ruleVersion = "string"
  ),
  description = "string",
  expression = "string",
  language = "DETECTORPL",
  outcomes = list(
    "string"
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
