% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_service.R
\name{fsx}
\alias{fsx}
\title{Amazon FSx}
\usage{
fsx(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon FSx is a fully managed service that makes it easy for storage and
application administrators to launch and use shared file storage.
}
\section{Service syntax}{
\preformatted{svc <- fsx(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=fsx_create_backup]{create_backup} \tab Creates a backup of an existing Amazon FSx for Windows File Server file system \cr
\link[=fsx_create_file_system]{create_file_system} \tab Creates a new, empty Amazon FSx file system \cr
\link[=fsx_create_file_system_from_backup]{create_file_system_from_backup} \tab Creates a new Amazon FSx file system from an existing Amazon FSx for Windows File Server backup \cr
\link[=fsx_delete_backup]{delete_backup} \tab Deletes an Amazon FSx for Windows File Server backup, deleting its contents \cr
\link[=fsx_delete_file_system]{delete_file_system} \tab Deletes a file system, deleting its contents \cr
\link[=fsx_describe_backups]{describe_backups} \tab Returns the description of specific Amazon FSx for Windows File Server backups, if a BackupIds value is provided for that backup\cr
\link[=fsx_describe_file_systems]{describe_file_systems} \tab Returns the description of specific Amazon FSx file systems, if a FileSystemIds value is provided for that file system \cr
\link[=fsx_list_tags_for_resource]{list_tags_for_resource} \tab Lists tags for an Amazon FSx file systems and backups in the case of Amazon FSx for Windows File Server \cr
\link[=fsx_tag_resource]{tag_resource} \tab Tags an Amazon FSx resource \cr
\link[=fsx_untag_resource]{untag_resource} \tab This action removes a tag from an Amazon FSx resource \cr
\link[=fsx_update_file_system]{update_file_system} \tab Updates a file system configuration
}
}

\examples{
# This operation creates a new backup.
\donttest{svc <- fsx()
svc$create_backup(
  FileSystemId = "fs-0498eed5fe91001ec",
  Tags = list(
    list(
      Key = "Name",
      Value = "MyBackup"
    )
  )
)}

}
