% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_build}
\alias{gamelift_delete_build}
\title{Deletes a build}
\usage{
gamelift_delete_build(BuildId)
}
\arguments{
\item{BuildId}{[required] A unique identifier for a build to delete. You can use either the build
ID or ARN value.}
}
\description{
Deletes a build. This operation permanently deletes the build resource
and any uploaded build files. Deleting a build does not affect the
status of any active fleets using the build, but you can no longer
create new fleets with the deleted build.
}
\details{
To delete a build, specify the build ID.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-build-intro.html}{Upload a Custom Server Build}

\strong{Related operations}
\itemize{
\item CreateBuild
\item ListBuilds
\item DescribeBuild
\item UpdateBuild
\item DeleteBuild
}
}
\section{Request syntax}{
\preformatted{svc$delete_build(
  BuildId = "string"
)
}
}

\keyword{internal}
