% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_alias}
\alias{gamelift_describe_alias}
\title{Retrieves properties for an alias}
\usage{
gamelift_describe_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] The unique identifier for the fleet alias that you want to retrieve. You
can use either the alias ID or ARN value.}
}
\description{
Retrieves properties for an alias. This operation returns all alias
metadata and settings. To get an alias's target fleet ID only, use
\code{ResolveAlias}.

To get alias properties, specify the alias ID. If successful, the
requested alias record is returned.
\itemize{
\item CreateAlias
\item ListAliases
\item DescribeAlias
\item UpdateAlias
\item DeleteAlias
\item ResolveAlias
}
}
\section{Request syntax}{
\preformatted{svc$describe_alias(
  AliasId = "string"
)
}
}

\keyword{internal}
