% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_capacity}
\alias{gamelift_describe_fleet_capacity}
\title{Retrieves the current capacity statistics for one or more fleets}
\usage{
gamelift_describe_fleet_capacity(FleetIds, Limit, NextToken)
}
\arguments{
\item{FleetIds}{A unique identifier for a fleet(s) to retrieve capacity information for.
You can use either the fleet ID or ARN value.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages. This parameter
is ignored when the request specifies one or a list of fleet IDs.}

\item{NextToken}{Token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this action. To
start at the beginning of the result set, do not specify a value. This
parameter is ignored when the request specifies one or a list of fleet
IDs.}
}
\description{
Retrieves the current capacity statistics for one or more fleets. These
statistics present a snapshot of the fleet's instances and provide
insight on current or imminent scaling activity. To get statistics on
game hosting activity in the fleet, see DescribeFleetUtilization.
}
\details{
You can request capacity for all fleets or specify a list of one or more
fleet identifiers. When requesting multiple fleets, use the pagination
parameters to retrieve results as a set of sequential pages. If
successful, a FleetCapacity object is returned for each requested fleet
ID. When a list of fleet IDs is provided, attribute objects are returned
only for fleets that currently exist.

Some API actions may limit the number of fleet IDs allowed in one
request. If a request exceeds this limit, the request fails and the
error message includes the maximum allowed.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/monitoring-cloudwatch.html#gamelift-metrics-fleet}{GameLift Metrics for Fleets}

\strong{Related operations}
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\itemize{
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
}
\item UpdateFleetAttributes
\item StartFleetActions or StopFleetActions
}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_capacity(
  FleetIds = list(
    "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
