% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_list_game_servers}
\alias{gamelift_list_game_servers}
\title{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups}
\usage{
gamelift_list_game_servers(GameServerGroupName, SortOrder, Limit,
  NextToken)
}
\arguments{
\item{GameServerGroupName}{[required] An identifier for the game server group to retrieve a list of game
servers from. Use either the GameServerGroup name or ARN value.}

\item{SortOrder}{Indicates how to sort the returned data based on game server
registration timestamp. Use ASCENDING to retrieve oldest game servers
first, or use DESCENDING to retrieve newest game servers first. If this
parameter is left empty, game servers are returned in no particular
order.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential segments.}

\item{NextToken}{A token that indicates the start of the next sequential segment of
results. Use the token returned with the previous call to this
operation. To start at the beginning of the result set, do not specify a
value.}
}
\description{
\strong{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups.}

Retrieves information on all game servers that are currently active in a
specified game server group. You can opt to sort the list by game server
age. Use the pagination parameters to retrieve results in a set of
sequential segments.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item RegisterGameServer
\item ListGameServers
\item ClaimGameServer
\item DescribeGameServer
\item UpdateGameServer
\item DeregisterGameServer
}
}
\section{Request syntax}{
\preformatted{svc$list_game_servers(
  GameServerGroupName = "string",
  SortOrder = "ASCENDING"|"DESCENDING",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
