% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_stop_game_session_placement}
\alias{gamelift_stop_game_session_placement}
\title{Cancels a game session placement that is in PENDING status}
\usage{
gamelift_stop_game_session_placement(PlacementId)
}
\arguments{
\item{PlacementId}{[required] A unique identifier for a game session placement to cancel.}
}
\description{
Cancels a game session placement that is in \code{PENDING} status. To stop a
placement, provide the placement ID values. If successful, the placement
is moved to \code{CANCELLED} status.
\itemize{
\item CreateGameSession
\item DescribeGameSessions
\item DescribeGameSessionDetails
\item SearchGameSessions
\item UpdateGameSession
\item GetGameSessionLogUrl
\item Game session placements
\itemize{
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
}
\section{Request syntax}{
\preformatted{svc$stop_game_session_placement(
  PlacementId = "string"
)
}
}

\keyword{internal}
