% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_suspend_game_server_group}
\alias{gamelift_suspend_game_server_group}
\title{THIS ACTION IS PART OF AMAZON GAMELIFT FLEETIQ WITH GAME SERVER GROUPS,
WHICH IS IN PREVIEW RELEASE AND IS SUBJECT TO CHANGE}
\usage{
gamelift_suspend_game_server_group(GameServerGroupName, SuspendActions)
}
\arguments{
\item{GameServerGroupName}{[required] The unique identifier of the game server group to stop activity on. Use
either the GameServerGroup name or ARN value.}

\item{SuspendActions}{[required] The action to suspend for this game server group.}
}
\description{
\strong{This action is part of Amazon GameLift FleetIQ with game server
groups, which is in preview release and is subject to change.}
}
\details{
Temporarily stops activity on a game server group without terminating
instances or the game server group. Activity can be restarted by calling
ResumeGameServerGroup. Activities that can suspended are:
\itemize{
\item Instance type replacement. This activity evaluates the current Spot
viability of all instance types that are defined for the game server
group. It updates the Auto Scaling group to remove nonviable Spot
instance types (which have a higher chance of game server
interruptions) and rebalances capacity across the remaining viable
Spot instance types. When this activity is suspended, the Auto
Scaling group continues with its current balance, regardless of
viability. Instance protection, utilization metrics, and capacity
autoscaling activities continue to be active.
}

To suspend activity, specify a game server group ARN and the type of
activity to be suspended.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item CreateGameServerGroup
\item ListGameServerGroups
\item DescribeGameServerGroup
\item UpdateGameServerGroup
\item DeleteGameServerGroup
\item ResumeGameServerGroup
\item SuspendGameServerGroup
}
}
\section{Request syntax}{
\preformatted{svc$suspend_game_server_group(
  GameServerGroupName = "string",
  SuspendActions = list(
    "REPLACE_INSTANCE_TYPES"
  )
)
}
}

\keyword{internal}
