% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_game_server}
\alias{gamelift_update_game_server}
\title{This action is part of Amazon GameLift FleetIQ with game server groups,
which is in preview release and is subject to change}
\usage{
gamelift_update_game_server(GameServerGroupName, GameServerId,
  GameServerData, CustomSortKey, UtilizationStatus, HealthCheck)
}
\arguments{
\item{GameServerGroupName}{[required] An identifier for the game server group where the game server is
running. Use either the GameServerGroup name or ARN value.}

\item{GameServerId}{[required] The identifier for the game server to be updated.}

\item{GameServerData}{A set of custom game server properties, formatted as a single string
value. This data is passed to a game client or service when it requests
information on a game servers using DescribeGameServer or
ClaimGameServer.}

\item{CustomSortKey}{A game server tag that can be used to request sorted lists of game
servers using ListGameServers. Custom sort keys are developer-defined
based on how you want to organize the retrieved game server information.}

\item{UtilizationStatus}{Indicates whether the game server is available or is currently hosting
gameplay.}

\item{HealthCheck}{Indicates health status of the game server. An update that explicitly
includes this parameter updates the game server's \emph{LastHealthCheckTime}
time stamp.}
}
\description{
\strong{This action is part of Amazon GameLift FleetIQ with game server
groups, which is in preview release and is subject to change.}
}
\details{
Updates information about a registered game server. This action is
called by a game server process that is running on an instance in a game
server group. There are three reasons to update game server information:
(1) to change the utilization status of the game server, (2) to report
game server health status, and (3) to change game server metadata. A
registered game server should regularly report health and should update
utilization status when it is supporting gameplay so that GameLift
FleetIQ can accurately track game server availability. You can make all
three types of updates in the same request.
\itemize{
\item To update the game server's utilization status, identify the game
server and game server group and specify the current utilization
status. Use this status to identify when game servers are currently
hosting games and when they are available to be claimed.
\item To report health status, identify the game server and game server
group and set health check to HEALTHY. If a game server does not
report health status for a certain length of time, the game server
is no longer considered healthy and will be eventually de-registered
from the game server group to avoid affecting utilization metrics.
The best practice is to report health every 60 seconds.
\item To change game server metadata, provide updated game server data and
custom sort key values.
}

Once a game server is successfully updated, the relevant statuses and
timestamps are updated.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item RegisterGameServer
\item ListGameServers
\item ClaimGameServer
\item DescribeGameServer
\item UpdateGameServer
\item DeregisterGameServer
}
}
\section{Request syntax}{
\preformatted{svc$update_game_server(
  GameServerGroupName = "string",
  GameServerId = "string",
  GameServerData = "string",
  CustomSortKey = "string",
  UtilizationStatus = "AVAILABLE"|"UTILIZED",
  HealthCheck = "HEALTHY"
)
}
}

\keyword{internal}
