% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_threat_intel_set}
\alias{guardduty_delete_threat_intel_set}
\title{Deletes ThreatIntelSet specified by the ThreatIntelSet ID}
\usage{
guardduty_delete_threat_intel_set(DetectorId, ThreatIntelSetId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector the threatIntelSet is associated with.}

\item{ThreatIntelSetId}{[required] The unique ID of the threatIntelSet you want to delete.}
}
\description{
Deletes ThreatIntelSet specified by the ThreatIntelSet ID.
}
\section{Request syntax}{
\preformatted{svc$delete_threat_intel_set(
  DetectorId = "string",
  ThreatIntelSetId = "string"
)
}
}

\keyword{internal}
