% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{health}
\alias{health}
\title{AWS Health APIs and Notifications}
\usage{
health(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{access_key_id}:} {AWS access key ID}
\item{\strong{secret_access_key}:} {AWS secret access key}
\item{\strong{session_token}:} {AWS temporary session token}
\item{\strong{profile}:} {The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}:} {Set anonymous credentials.}
\item{\strong{endpoint}:} {The complete URL to use for the constructed client.}
\item{\strong{region}:} {The AWS Region used in instantiating the client.}
\item{\strong{close_connection}:} {Immediately close all HTTP connections.}
\item{\strong{timeout}:} {The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}:} {Set this to \code{true} to force the request to use path-style addressing, i.e., \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
}}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Health

The Health API provides access to the Health information that appears in
the \href{https://health.aws.amazon.com/health/home}{Health Dashboard}. You
can use the API operations to get information about events that might
affect your Amazon Web Services and resources.

You must have a Business, Enterprise On-Ramp, or Enterprise Support plan
from \href{https://aws.amazon.com/premiumsupport/}{Amazon Web Services Support} to use the Health API.
If you call the Health API from an Amazon Web Services account that
doesn't have a Business, Enterprise On-Ramp, or Enterprise Support plan,
you receive a \code{SubscriptionRequiredException} error.

For API access, you need an access key ID and a secret access key. Use
temporary credentials instead of long-term access keys when possible.
Temporary credentials include an access key ID, a secret access key, and
a security token that indicates when the credentials expire. For more
information, see \href{https://docs.aws.amazon.com/accounts/latest/reference/best-practices.html}{Best practices for managing Amazon Web Services access keys}
in the \emph{Amazon Web Services General Reference}.

You can use the Health endpoint health.us-east-1.amazonaws.com (HTTPS)
to call the Health API operations. Health supports a multi-Region
application architecture and has two regional endpoints in an
active-passive configuration. You can use the high availability endpoint
example to determine which Amazon Web Services Region is active, so that
you can get the latest information from the API. For more information,
see \href{https://docs.aws.amazon.com/health/latest/ug/health-api.html}{Accessing the Health API} in
the \emph{Health User Guide}.

For authentication of requests, Health uses the \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_aws-signing.html}{Signature Version 4 Signing Process}.

If your Amazon Web Services account is part of Organizations, you can
use the Health organizational view feature. This feature provides a
centralized view of Health events across all accounts in your
organization. You can aggregate Health events in real time to identify
accounts in your organization that are affected by an operational event
or get notified of security vulnerabilities. Use the organizational view
API operations to enable this feature and return event information. For
more information, see \href{https://docs.aws.amazon.com/health/latest/ug/aggregate-events.html}{Aggregating Health events}
in the \emph{Health User Guide}.

When you use the Health API operations to return Health events, see the
following recommendations:
\itemize{
\item Use the
\href{https://docs.aws.amazon.com/health/latest/APIReference/API_Event.html#AWSHealth-Type-Event-eventScopeCode}{eventScopeCode}
parameter to specify whether to return Health events that are public
or account-specific.
\item Use pagination to view all events from the response. For example, if
you call the
\code{describe_events_for_organization}
operation to get all events in your organization, you might receive
several page results. Specify the \code{nextToken} in the next request to
return more results.
}
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- health(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical"
  )
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.management:health_describe_affected_accounts_for_organization]{describe_affected_accounts_for_organization} \tab Returns a list of accounts in the organization from Organizations that are affected by the provided event\cr
\link[paws.management:health_describe_affected_entities]{describe_affected_entities} \tab Returns a list of entities that have been affected by the specified events, based on the specified filter criteria\cr
\link[paws.management:health_describe_affected_entities_for_organization]{describe_affected_entities_for_organization} \tab Returns a list of entities that have been affected by one or more events for one or more accounts in your organization in Organizations, based on the filter criteria\cr
\link[paws.management:health_describe_entity_aggregates]{describe_entity_aggregates} \tab Returns the number of entities that are affected by each of the specified events\cr
\link[paws.management:health_describe_event_aggregates]{describe_event_aggregates} \tab Returns the number of events of each event type (issue, scheduled change, and account notification)\cr
\link[paws.management:health_describe_event_details]{describe_event_details} \tab Returns detailed information about one or more specified events\cr
\link[paws.management:health_describe_event_details_for_organization]{describe_event_details_for_organization} \tab Returns detailed information about one or more specified events for one or more Amazon Web Services accounts in your organization\cr
\link[paws.management:health_describe_events]{describe_events} \tab Returns information about events that meet the specified filter criteria\cr
\link[paws.management:health_describe_events_for_organization]{describe_events_for_organization} \tab Returns information about events across your organization in Organizations\cr
\link[paws.management:health_describe_event_types]{describe_event_types} \tab Returns the event types that meet the specified filter criteria\cr
\link[paws.management:health_describe_health_service_status_for_organization]{describe_health_service_status_for_organization} \tab This operation provides status information on enabling or disabling Health to work with your organization\cr
\link[paws.management:health_disable_health_service_access_for_organization]{disable_health_service_access_for_organization} \tab Disables Health from working with Organizations\cr
\link[paws.management:health_enable_health_service_access_for_organization]{enable_health_service_access_for_organization} \tab Enables Health to work with Organizations
}
}

\examples{
\dontrun{
svc <- health()
svc$describe_affected_accounts_for_organization(
  Foo = 123
)
}

}
