% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_service.R
\name{healthlake}
\alias{healthlake}
\title{Amazon HealthLake}
\usage{
healthlake(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon HealthLake is a HIPAA eligibile service that allows customers to
store, transform, query, and analyze their FHIR-formatted data in a
consistent fashion in the cloud.
}
\section{Service syntax}{
\preformatted{svc <- healthlake(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=healthlake_create_fhir_datastore]{create_fhir_datastore} \tab Creates a Data Store that can ingest and export FHIR formatted data\cr
\link[=healthlake_delete_fhir_datastore]{delete_fhir_datastore} \tab Deletes a Data Store\cr
\link[=healthlake_describe_fhir_datastore]{describe_fhir_datastore} \tab Gets the properties associated with the FHIR Data Store, including the Data Store ID, Data Store ARN, Data Store name, Data Store status, created at, Data Store type version, and Data Store endpoint\cr
\link[=healthlake_describe_fhir_export_job]{describe_fhir_export_job} \tab Displays the properties of a FHIR export job, including the ID, ARN, name, and the status of the job\cr
\link[=healthlake_describe_fhir_import_job]{describe_fhir_import_job} \tab Displays the properties of a FHIR import job, including the ID, ARN, name, and the status of the job\cr
\link[=healthlake_list_fhir_datastores]{list_fhir_datastores} \tab Lists all FHIR Data Stores that are in the user’s account, regardless of Data Store status\cr
\link[=healthlake_start_fhir_export_job]{start_fhir_export_job} \tab Begins a FHIR export job\cr
\link[=healthlake_start_fhir_import_job]{start_fhir_import_job} \tab Begins a FHIR Import job
}
}

\examples{
\dontrun{
svc <- healthlake()
svc$create_fhir_datastore(
  Foo = 123
)
}

}
