% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycode_service.R
\name{honeycode}
\alias{honeycode}
\title{Amazon Honeycode}
\usage{
honeycode(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Honeycode is a fully managed service that allows you to quickly
build mobile and web apps for teams—without programming. Build Honeycode
apps for managing almost anything, like projects, customers, operations,
approvals, resources, and even your team.
}
\section{Service syntax}{
\preformatted{svc <- honeycode(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=honeycode_get_screen_data]{get_screen_data} \tab The GetScreenData API allows retrieval of data from a screen in a Honeycode app\cr
\link[=honeycode_invoke_screen_automation]{invoke_screen_automation} \tab The InvokeScreenAutomation API allows invoking an action defined in a screen in a Honeycode app
}
}

\examples{
\dontrun{
svc <- honeycode()
svc$get_screen_data(
  Foo = 123
)
}

}
