% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_create_image}
\alias{imagebuilder_create_image}
\title{Creates a new image}
\usage{
imagebuilder_create_image(imageRecipeArn, containerRecipeArn,
  distributionConfigurationArn, infrastructureConfigurationArn,
  imageTestsConfiguration, enhancedImageMetadataEnabled, tags,
  clientToken)
}
\arguments{
\item{imageRecipeArn}{The Amazon Resource Name (ARN) of the image recipe that defines how
images are configured, tested, and assessed.}

\item{containerRecipeArn}{The Amazon Resource Name (ARN) of the container recipe that defines how
images are configured and tested.}

\item{distributionConfigurationArn}{The Amazon Resource Name (ARN) of the distribution configuration that
defines and configures the outputs of your pipeline.}

\item{infrastructureConfigurationArn}{[required] The Amazon Resource Name (ARN) of the infrastructure configuration that
defines the environment in which your image will be built and tested.}

\item{imageTestsConfiguration}{The image tests configuration of the image.}

\item{enhancedImageMetadataEnabled}{Collects additional information about the image being created, including
the operating system (OS) version and package list. This information is
used to enhance the overall experience of using EC2 Image Builder.
Enabled by default.}

\item{tags}{The tags of the image.}

\item{clientToken}{[required] The idempotency token used to make this request idempotent.}
}
\description{
Creates a new image. This request will create a new image along with all
of the configured output resources defined in the distribution
configuration.
}
\section{Request syntax}{
\preformatted{svc$create_image(
  imageRecipeArn = "string",
  containerRecipeArn = "string",
  distributionConfigurationArn = "string",
  infrastructureConfigurationArn = "string",
  imageTestsConfiguration = list(
    imageTestsEnabled = TRUE|FALSE,
    timeoutMinutes = 123
  ),
  enhancedImageMetadataEnabled = TRUE|FALSE,
  tags = list(
    "string"
  ),
  clientToken = "string"
)
}
}

\keyword{internal}
