% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_create_image_recipe}
\alias{imagebuilder_create_image_recipe}
\title{Creates a new image recipe}
\usage{
imagebuilder_create_image_recipe(name, description, semanticVersion,
  components, parentImage, blockDeviceMappings, tags, workingDirectory,
  clientToken)
}
\arguments{
\item{name}{[required] The name of the image recipe.}

\item{description}{The description of the image recipe.}

\item{semanticVersion}{[required] The semantic version of the image recipe.}

\item{components}{[required] The components of the image recipe.}

\item{parentImage}{[required] The parent image of the image recipe. The value of the string can be the
ARN of the parent image or an AMI ID. The format for the ARN follows
this example:
\code{arn:aws:imagebuilder:us-west-2:aws:image/windows-server-2016-english-full-base-x86/xxxx.x.x}.
You can provide the specific version that you want to use, or you can
use a wildcard in all of the fields. If you enter an AMI ID for the
string value, you must have access to the AMI, and the AMI must be in
the same Region in which you are using Image Builder.}

\item{blockDeviceMappings}{The block device mappings of the image recipe.}

\item{tags}{The tags of the image recipe.}

\item{workingDirectory}{The working directory to be used during build and test workflows.}

\item{clientToken}{[required] The idempotency token used to make this request idempotent.}
}
\description{
Creates a new image recipe. Image recipes define how images are
configured, tested, and assessed.
}
\section{Request syntax}{
\preformatted{svc$create_image_recipe(
  name = "string",
  description = "string",
  semanticVersion = "string",
  components = list(
    list(
      componentArn = "string"
    )
  ),
  parentImage = "string",
  blockDeviceMappings = list(
    list(
      deviceName = "string",
      ebs = list(
        encrypted = TRUE|FALSE,
        deleteOnTermination = TRUE|FALSE,
        iops = 123,
        kmsKeyId = "string",
        snapshotId = "string",
        volumeSize = 123,
        volumeType = "standard"|"io1"|"io2"|"gp2"|"sc1"|"st1"
      ),
      virtualName = "string",
      noDevice = "string"
    )
  ),
  tags = list(
    "string"
  ),
  workingDirectory = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
