% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_build_versions}
\alias{imagebuilder_list_image_build_versions}
\title{Returns a list of distribution configurations}
\usage{
imagebuilder_list_image_build_versions(imageVersionArn, filters,
  maxResults, nextToken)
}
\arguments{
\item{imageVersionArn}{[required] The Amazon Resource Name (ARN) of the image whose build versions you
wish to retrieve.}

\item{filters}{}

\item{maxResults}{The maximum items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the NextToken from
a previously truncated response.}
}
\description{
Returns a list of distribution configurations.
}
\section{Request syntax}{
\preformatted{svc$list_image_build_versions(
  imageVersionArn = "string",
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
