% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_images}
\alias{imagebuilder_list_images}
\title{Returns the list of images that you have access to}
\usage{
imagebuilder_list_images(owner, filters, maxResults, nextToken)
}
\arguments{
\item{owner}{The owner defines which images you want to list. By default, this
request will only show images owned by your account. You can use this
field to specify if you want to view images owned by yourself, by
Amazon, or those images that have been shared with you by other
customers.}

\item{filters}{The filters.}

\item{maxResults}{The maximum items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the NextToken from
a previously truncated response.}
}
\description{
Returns the list of images that you have access to.
}
\section{Request syntax}{
\preformatted{svc$list_images(
  owner = "Self"|"Shared"|"Amazon",
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
