% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_finalize_device_claim}
\alias{iot1clickdevicesservice_finalize_device_claim}
\title{Given a device ID, finalizes the claim request for the associated device}
\usage{
iot1clickdevicesservice_finalize_device_claim(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\description{
Given a device ID, finalizes the claim request for the associated
device.
}
\details{
Claiming a device consists of initiating a claim, then publishing a
device event, and finalizing the claim. For a device of type button, a
device event can be published by simply clicking the device.
}
\section{Request syntax}{
\preformatted{iot1clickdevicesservice$finalize_device_claim(
  DeviceId = "string"
)
}
}

\keyword{internal}
