% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_untag_resource}
\alias{iot1clickdevicesservice_untag_resource}
\title{Using tag keys, deletes the tags (key/value pairs) associated with the
specified resource ARN}
\usage{
iot1clickdevicesservice_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource.}

\item{TagKeys}{[required] A collections of tag keys. For example, \{"key1","key2"\}}
}
\description{
Using tag keys, deletes the tags (key/value pairs) associated with the
specified resource ARN.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
