% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_create_project}
\alias{iot1clickprojects_create_project}
\title{Creates an empty project with a placement template}
\usage{
iot1clickprojects_create_project(projectName, description,
  placementTemplate, tags)
}
\arguments{
\item{projectName}{[required] The name of the project to create.}

\item{description}{An optional description for the project.}

\item{placementTemplate}{The schema defining the placement to be created. A placement template
defines placement default attributes and device templates. You cannot
add or remove device templates after the project has been created.
However, you can update \code{callbackOverrides} for the device templates
using the \code{\link[=iot1clickprojects_update_project]{update_project}} API.}

\item{tags}{Optional tags (metadata key/value pairs) to be associated with the
project. For example, \verb{\\\{ \\\{"key1": "value1", "key2": "value2"\\\} \\\}}. For
more information, see \href{https://d1.awsstatic.com/whitepapers/aws-tagging-best-practices.pdf}{AWS Tagging Strategies}.}
}
\description{
Creates an empty project with a placement template. A project contains
zero or more placements that adhere to the placement template defined in
the project.
}
\section{Request syntax}{
\preformatted{svc$create_project(
  projectName = "string",
  description = "string",
  placementTemplate = list(
    defaultAttributes = list(
      "string"
    ),
    deviceTemplates = list(
      list(
        deviceType = "string",
        callbackOverrides = list(
          "string"
        )
      )
    )
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
