% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_register_thing}
\alias{iot_register_thing}
\title{Provisions a thing in the device registry}
\usage{
iot_register_thing(templateBody, parameters)
}
\arguments{
\item{templateBody}{[required] The provisioning template. See \href{https://docs.aws.amazon.com/iot/latest/developerguide/programmatic-provisioning.html}{Programmatic Provisioning}
for more information.}

\item{parameters}{The parameters for provisioning a thing. See \href{https://docs.aws.amazon.com/iot/latest/developerguide/programmatic-provisioning.html}{Programmatic Provisioning}
for more information.}
}
\description{
Provisions a thing in the device registry. RegisterThing calls other AWS
IoT control plane APIs. These calls might exceed your account level \href{https://docs.aws.amazon.com/general/latest/gr/aws_service_limits.html#limits_iot}{AWS IoT Throttling Limits}
and cause throttle errors. Please contact \href{https://console.aws.amazon.com/support/home}{AWS Customer Support} to raise your
throttling limits if necessary.
}
\section{Request syntax}{
\preformatted{svc$register_thing(
  templateBody = "string",
  parameters = list(
    "string"
  )
)
}
}

\keyword{internal}
