% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_create_datastore}
\alias{iotanalytics_create_datastore}
\title{Creates a data store, which is a repository for messages}
\usage{
iotanalytics_create_datastore(datastoreName, retentionPeriod, tags)
}
\arguments{
\item{datastoreName}{[required] The name of the data store.}

\item{retentionPeriod}{How long, in days, message data is kept for the data store.}

\item{tags}{Metadata which can be used to manage the data store.}
}
\description{
Creates a data store, which is a repository for messages.
}
\section{Request syntax}{
\preformatted{svc$create_datastore(
  datastoreName = "string",
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
