% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_describe_channel}
\alias{iotanalytics_describe_channel}
\title{Retrieves information about a channel}
\usage{
iotanalytics_describe_channel(channelName, includeStatistics)
}
\arguments{
\item{channelName}{[required] The name of the channel whose information is retrieved.}

\item{includeStatistics}{If true, additional statistical information about the channel is
included in the response. This feature cannot be used with a channel
whose S3 storage is customer-managed.}
}
\description{
Retrieves information about a channel.
}
\section{Request syntax}{
\preformatted{svc$describe_channel(
  channelName = "string",
  includeStatistics = TRUE|FALSE
)
}
}

\keyword{internal}
