% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_create_detector_model}
\alias{iotevents_create_detector_model}
\title{Creates a detector model}
\usage{
iotevents_create_detector_model(detectorModelName,
  detectorModelDefinition, detectorModelDescription, key, roleArn, tags,
  evaluationMethod)
}
\arguments{
\item{detectorModelName}{[required] The name of the detector model.}

\item{detectorModelDefinition}{[required] Information that defines how the detectors operate.}

\item{detectorModelDescription}{A brief description of the detector model.}

\item{key}{The input attribute key used to identify a device or system to create a
detector (an instance of the detector model) and then to route each
input received to the appropriate detector (instance). This parameter
uses a JSON-path expression to specify the attribute-value pair in the
message payload of each input that is used to identify the device
associated with the input.}

\item{roleArn}{[required] The ARN of the role that grants permission to AWS IoT Events to perform
its operations.}

\item{tags}{Metadata that can be used to manage the detector model.}

\item{evaluationMethod}{When set to \code{SERIAL}, variables are updated and event conditions
evaluated in the order that the events are defined. When set to \code{BATCH},
variables are updated and events performed only after all event
conditions are evaluated.}
}
\description{
Creates a detector model.
}
\section{Request syntax}{
\preformatted{svc$create_detector_model(
  detectorModelName = "string",
  detectorModelDefinition = list(
    states = list(
      list(
        stateName = "string",
        onInput = list(
          events = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string"
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string"
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  ),
                  lambda = list(
                    functionArn = "string"
                  ),
                  iotEvents = list(
                    inputName = "string"
                  ),
                  sqs = list(
                    queueUrl = "string",
                    useBase64 = TRUE|FALSE
                  ),
                  firehose = list(
                    deliveryStreamName = "string",
                    separator = "string"
                  )
                )
              )
            )
          ),
          transitionEvents = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string"
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string"
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  ),
                  lambda = list(
                    functionArn = "string"
                  ),
                  iotEvents = list(
                    inputName = "string"
                  ),
                  sqs = list(
                    queueUrl = "string",
                    useBase64 = TRUE|FALSE
                  ),
                  firehose = list(
                    deliveryStreamName = "string",
                    separator = "string"
                  )
                )
              ),
              nextState = "string"
            )
          )
        ),
        onEnter = list(
          events = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string"
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string"
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  ),
                  lambda = list(
                    functionArn = "string"
                  ),
                  iotEvents = list(
                    inputName = "string"
                  ),
                  sqs = list(
                    queueUrl = "string",
                    useBase64 = TRUE|FALSE
                  ),
                  firehose = list(
                    deliveryStreamName = "string",
                    separator = "string"
                  )
                )
              )
            )
          )
        ),
        onExit = list(
          events = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string"
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string"
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  ),
                  lambda = list(
                    functionArn = "string"
                  ),
                  iotEvents = list(
                    inputName = "string"
                  ),
                  sqs = list(
                    queueUrl = "string",
                    useBase64 = TRUE|FALSE
                  ),
                  firehose = list(
                    deliveryStreamName = "string",
                    separator = "string"
                  )
                )
              )
            )
          )
        )
      )
    ),
    initialStateName = "string"
  ),
  detectorModelDescription = "string",
  key = "string",
  roleArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  ),
  evaluationMethod = "BATCH"|"SERIAL"
)
}
}

\keyword{internal}
