% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotevents_operations.R
\name{iotevents_update_detector_model}
\alias{iotevents_update_detector_model}
\title{Updates a detector model}
\usage{
iotevents_update_detector_model(detectorModelName,
  detectorModelDefinition, detectorModelDescription, roleArn,
  evaluationMethod)
}
\arguments{
\item{detectorModelName}{[required] The name of the detector model that is updated.}

\item{detectorModelDefinition}{[required] Information that defines how a detector operates.}

\item{detectorModelDescription}{A brief description of the detector model.}

\item{roleArn}{[required] The ARN of the role that grants permission to AWS IoT Events to perform
its operations.}

\item{evaluationMethod}{Information about the order in which events are evaluated and how
actions are executed.}
}
\description{
Updates a detector model. Detectors (instances) spawned by the previous
version are deleted and then re-created as new inputs arrive.
}
\section{Request syntax}{
\preformatted{svc$update_detector_model(
  detectorModelName = "string",
  detectorModelDefinition = list(
    states = list(
      list(
        stateName = "string",
        onInput = list(
          events = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123,
                    durationExpression = "string"
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  ),
                  lambda = list(
                    functionArn = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotEvents = list(
                    inputName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  sqs = list(
                    queueUrl = "string",
                    useBase64 = TRUE|FALSE,
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  firehose = list(
                    deliveryStreamName = "string",
                    separator = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  dynamoDB = list(
                    hashKeyType = "string",
                    hashKeyField = "string",
                    hashKeyValue = "string",
                    rangeKeyType = "string",
                    rangeKeyField = "string",
                    rangeKeyValue = "string",
                    operation = "string",
                    payloadField = "string",
                    tableName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  dynamoDBv2 = list(
                    tableName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotSiteWise = list(
                    entryId = "string",
                    assetId = "string",
                    propertyId = "string",
                    propertyAlias = "string",
                    propertyValue = list(
                      value = list(
                        stringValue = "string",
                        integerValue = "string",
                        doubleValue = "string",
                        booleanValue = "string"
                      ),
                      timestamp = list(
                        timeInSeconds = "string",
                        offsetInNanos = "string"
                      ),
                      quality = "string"
                    )
                  )
                )
              )
            )
          ),
          transitionEvents = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123,
                    durationExpression = "string"
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  ),
                  lambda = list(
                    functionArn = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotEvents = list(
                    inputName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  sqs = list(
                    queueUrl = "string",
                    useBase64 = TRUE|FALSE,
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  firehose = list(
                    deliveryStreamName = "string",
                    separator = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  dynamoDB = list(
                    hashKeyType = "string",
                    hashKeyField = "string",
                    hashKeyValue = "string",
                    rangeKeyType = "string",
                    rangeKeyField = "string",
                    rangeKeyValue = "string",
                    operation = "string",
                    payloadField = "string",
                    tableName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  dynamoDBv2 = list(
                    tableName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotSiteWise = list(
                    entryId = "string",
                    assetId = "string",
                    propertyId = "string",
                    propertyAlias = "string",
                    propertyValue = list(
                      value = list(
                        stringValue = "string",
                        integerValue = "string",
                        doubleValue = "string",
                        booleanValue = "string"
                      ),
                      timestamp = list(
                        timeInSeconds = "string",
                        offsetInNanos = "string"
                      ),
                      quality = "string"
                    )
                  )
                )
              ),
              nextState = "string"
            )
          )
        ),
        onEnter = list(
          events = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123,
                    durationExpression = "string"
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  ),
                  lambda = list(
                    functionArn = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotEvents = list(
                    inputName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  sqs = list(
                    queueUrl = "string",
                    useBase64 = TRUE|FALSE,
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  firehose = list(
                    deliveryStreamName = "string",
                    separator = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  dynamoDB = list(
                    hashKeyType = "string",
                    hashKeyField = "string",
                    hashKeyValue = "string",
                    rangeKeyType = "string",
                    rangeKeyField = "string",
                    rangeKeyValue = "string",
                    operation = "string",
                    payloadField = "string",
                    tableName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  dynamoDBv2 = list(
                    tableName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotSiteWise = list(
                    entryId = "string",
                    assetId = "string",
                    propertyId = "string",
                    propertyAlias = "string",
                    propertyValue = list(
                      value = list(
                        stringValue = "string",
                        integerValue = "string",
                        doubleValue = "string",
                        booleanValue = "string"
                      ),
                      timestamp = list(
                        timeInSeconds = "string",
                        offsetInNanos = "string"
                      ),
                      quality = "string"
                    )
                  )
                )
              )
            )
          )
        ),
        onExit = list(
          events = list(
            list(
              eventName = "string",
              condition = "string",
              actions = list(
                list(
                  setVariable = list(
                    variableName = "string",
                    value = "string"
                  ),
                  sns = list(
                    targetArn = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotTopicPublish = list(
                    mqttTopic = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  setTimer = list(
                    timerName = "string",
                    seconds = 123,
                    durationExpression = "string"
                  ),
                  clearTimer = list(
                    timerName = "string"
                  ),
                  resetTimer = list(
                    timerName = "string"
                  ),
                  lambda = list(
                    functionArn = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotEvents = list(
                    inputName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  sqs = list(
                    queueUrl = "string",
                    useBase64 = TRUE|FALSE,
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  firehose = list(
                    deliveryStreamName = "string",
                    separator = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  dynamoDB = list(
                    hashKeyType = "string",
                    hashKeyField = "string",
                    hashKeyValue = "string",
                    rangeKeyType = "string",
                    rangeKeyField = "string",
                    rangeKeyValue = "string",
                    operation = "string",
                    payloadField = "string",
                    tableName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  dynamoDBv2 = list(
                    tableName = "string",
                    payload = list(
                      contentExpression = "string",
                      type = "STRING"|"JSON"
                    )
                  ),
                  iotSiteWise = list(
                    entryId = "string",
                    assetId = "string",
                    propertyId = "string",
                    propertyAlias = "string",
                    propertyValue = list(
                      value = list(
                        stringValue = "string",
                        integerValue = "string",
                        doubleValue = "string",
                        booleanValue = "string"
                      ),
                      timestamp = list(
                        timeInSeconds = "string",
                        offsetInNanos = "string"
                      ),
                      quality = "string"
                    )
                  )
                )
              )
            )
          )
        )
      )
    ),
    initialStateName = "string"
  ),
  detectorModelDescription = "string",
  roleArn = "string",
  evaluationMethod = "BATCH"|"SERIAL"
)
}
}

\keyword{internal}
