% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioteventsdata_service.R
\name{ioteventsdata}
\alias{ioteventsdata}
\title{AWS IoT Events Data}
\usage{
ioteventsdata(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
AWS IoT Events monitors your equipment or device fleets for failures or
changes in operation, and triggers actions when such events occur. AWS
IoT Events Data API commands enable you to send inputs to detectors,
list detectors, and view or update a detector\'s status.
}
\section{Service syntax}{
\preformatted{svc <- ioteventsdata(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=ioteventsdata_batch_put_message]{batch_put_message} \tab Sends a set of messages to the AWS IoT Events system \cr
\link[=ioteventsdata_batch_update_detector]{batch_update_detector} \tab Updates the state, variable values, and timer settings of one or more detectors (instances) of a specified detector model\cr
\link[=ioteventsdata_describe_detector]{describe_detector} \tab Returns information about the specified detector (instance) \cr
\link[=ioteventsdata_list_detectors]{list_detectors} \tab Lists detectors (the instances of a detector model)
}
}

\examples{
\dontrun{svc <- ioteventsdata()
svc$batch_put_message(
  Foo = 123
)}

}
