% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_create_portal}
\alias{iotsitewise_create_portal}
\title{Creates a portal, which can contain projects and dashboards}
\usage{
iotsitewise_create_portal(portalName, portalDescription,
  portalContactEmail, clientToken, portalLogoImageFile, roleArn, tags,
  portalAuthMode)
}
\arguments{
\item{portalName}{[required] A friendly name for the portal.}

\item{portalDescription}{A description for the portal.}

\item{portalContactEmail}{[required] The AWS administrator's contact email address.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don't reuse this client token if a new
idempotent request is required.}

\item{portalLogoImageFile}{A logo image to display in the portal. Upload a square, high-resolution
image. The image is displayed on a dark background.}

\item{roleArn}{[required] The
\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN}
of a service role that allows the portal's users to access your AWS IoT
SiteWise resources on your behalf. For more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/monitor-service-role.html}{Using service roles for AWS IoT SiteWise Monitor}
in the \emph{AWS IoT SiteWise User Guide}.}

\item{tags}{A list of key-value pairs that contain metadata for the portal. For more
information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/tag-resources.html}{Tagging your AWS IoT SiteWise resources}
in the \emph{AWS IoT SiteWise User Guide}.}

\item{portalAuthMode}{The service to use to authenticate users to the portal. Choose from the
following options:
\itemize{
\item \code{SSO} – The portal uses AWS Single Sign-On to authenticate users and
manage user permissions. Before you can create a portal that uses
AWS SSO, you must enable AWS SSO. For more information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/monitor-getting-started.html#mon-gs-sso}{Enabling AWS SSO}
in the \emph{AWS IoT SiteWise User Guide}. This option is only available
in AWS Regions other than the China Regions.
\item \code{IAM} – The portal uses AWS Identity and Access Management (IAM) to
authenticate users and manage user permissions. This option is only
available in the China Regions.
}

You can't change this value after you create a portal.

Default: \code{SSO}}
}
\description{
Creates a portal, which can contain projects and dashboards. AWS IoT
SiteWise Monitor uses AWS SSO or IAM to authenticate portal users and
manage user permissions.

Before you can sign in to a new portal, you must add at least one
identity to that portal. For more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/administer-portals.html#portal-change-admins}{Adding or removing portal administrators}
in the \emph{AWS IoT SiteWise User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_portal(
  portalName = "string",
  portalDescription = "string",
  portalContactEmail = "string",
  clientToken = "string",
  portalLogoImageFile = list(
    data = raw,
    type = "PNG"
  ),
  roleArn = "string",
  tags = list(
    "string"
  ),
  portalAuthMode = "IAM"|"SSO"
)
}
}

\keyword{internal}
