% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_delete_asset}
\alias{iotsitewise_delete_asset}
\title{Deletes an asset}
\usage{
iotsitewise_delete_asset(assetId, clientToken)
}
\arguments{
\item{assetId}{[required] The ID of the asset to delete.}

\item{clientToken}{A unique case-sensitive identifier that you can provide to ensure the
idempotency of the request. Don't reuse this client token if a new
idempotent request is required.}
}
\description{
Deletes an asset. This action can't be undone. For more information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/delete-assets-and-models.html}{Deleting assets and models}
in the \emph{AWS IoT SiteWise User Guide}.
}
\details{
You can't delete an asset that's associated to another asset. For more
information, see
\href{https://docs.aws.amazon.com/iot-sitewise/latest/APIReference/API_DisassociateAssets.html}{DisassociateAssets}.
}
\section{Request syntax}{
\preformatted{svc$delete_asset(
  assetId = "string",
  clientToken = "string"
)
}
}

\keyword{internal}
