% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_describe_asset_property}
\alias{iotsitewise_describe_asset_property}
\title{Retrieves information about an asset's property}
\usage{
iotsitewise_describe_asset_property(assetId, propertyId)
}
\arguments{
\item{assetId}{[required] The ID of the asset.}

\item{propertyId}{[required] The ID of the asset property.}
}
\description{
Retrieves information about an asset's property.
}
\section{Request syntax}{
\preformatted{svc$describe_asset_property(
  assetId = "string",
  propertyId = "string"
)
}
}

\keyword{internal}
