% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotsitewise_operations.R
\name{iotsitewise_list_associated_assets}
\alias{iotsitewise_list_associated_assets}
\title{Retrieves a paginated list of associated assets}
\usage{
iotsitewise_list_associated_assets(assetId, hierarchyId,
  traversalDirection, nextToken, maxResults)
}
\arguments{
\item{assetId}{[required] The ID of the asset to query.}

\item{hierarchyId}{The ID of the hierarchy by which child assets are associated to the
asset. To find a hierarchy ID, use the
\code{\link[=iotsitewise_describe_asset]{describe_asset}} or
\code{\link[=iotsitewise_describe_asset_model]{describe_asset_model}} operations.
This parameter is required if you choose \code{CHILD} for
\code{traversalDirection}.

For more information, see \href{https://docs.aws.amazon.com/iot-sitewise/latest/userguide/asset-hierarchies.html}{Asset hierarchies}
in the \emph{AWS IoT SiteWise User Guide}.}

\item{traversalDirection}{The direction to list associated assets. Choose one of the following
options:
\itemize{
\item \code{CHILD} – The list includes all child assets associated to the
asset. The \code{hierarchyId} parameter is required if you choose
\code{CHILD}.
\item \code{PARENT} – The list includes the asset's parent asset.
}

Default: \code{CHILD}}

\item{nextToken}{The token to be used for the next set of paginated results.}

\item{maxResults}{The maximum number of results to be returned per paginated request.

Default: 50}
}
\description{
Retrieves a paginated list of associated assets.

You can use this operation to do the following:
\itemize{
\item List child assets associated to a parent asset by a hierarchy that
you specify.
\item List an asset's parent asset.
}
}
\section{Request syntax}{
\preformatted{svc$list_associated_assets(
  assetId = "string",
  hierarchyId = "string",
  traversalDirection = "PARENT"|"CHILD",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
