% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_service.R
\name{ivs}
\alias{ivs}
\title{Amazon Interactive Video Service}
\usage{
ivs(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
\strong{Introduction}

The Amazon Interactive Video Service (IVS) API is REST compatible, using
a standard HTTP API and an \href{http://aws.amazon.com/sns}{AWS SNS} event
stream for responses. JSON is used for both requests and responses,
including errors.

The API is an AWS regional service, currently in these regions:
us-west-2, us-east-2, and eu-west-1.

\emph{\strong{All API request parameters and URLs are case sensitive.}}

For a summary of notable documentation changes in each release, see
\href{https://docs.aws.amazon.com/ivs/latest/userguide/doc-history.html}{Document History}.

\strong{Allowed Header Values}
\itemize{
\item \verb{<b>Accept:</b>} application/json
\item \verb{<b>Accept-Encoding:</b>} gzip, deflate
\item \verb{<b>Content-Type:</b>}application/json
}

\strong{Resources}

The following resources contain information about your IVS live stream
(see \href{https://docs.aws.amazon.com/ivs/latest/userguide/GSIVS.html}{Getting Started with Amazon IVS}):
\itemize{
\item Channel — Stores configuration data related to your live stream. You
first create a channel and then use the channel’s stream key to
start your live stream. See the Channel endpoints for more
information.
\item Stream key — An identifier assigned by Amazon IVS when you create a
channel, which is then used to authorize streaming. See the
StreamKey endpoints for more information. \emph{\strong{Treat the stream key
like a secret, since it allows anyone to stream to the channel.}}
}

\strong{Tagging}

A \emph{tag} is a metadata label that you assign to an AWS resource. A tag
comprises a \emph{key} and a \emph{value}, both set by you. For example, you might
set a tag as \code{topic:nature} to label a particular video category. See
\href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS Resources}
for more information, including restrictions that apply to tags.

Tags can help you identify and organize your AWS resources. For example,
you can use the same tag for different resources to indicate that they
are related. You can also use tags to manage access (see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access_tags.html}{Access Tags}).

The Amazon IVS API has these tag-related endpoints: TagResource,
UntagResource, and ListTagsForResource. The following resources support
tagging: Channels and Stream Keys.

\strong{API Endpoints}

Channel:
\itemize{
\item CreateChannel — Creates a new channel and an associated stream key
to start streaming.
\item GetChannel — Gets the channel configuration for a specified channel
ARN (Amazon Resource Name).
\item BatchGetChannel — Performs GetChannel on multiple ARNs
simultaneously.
\item ListChannels — Gets summary information about channels. This list
can be filtered to match a specified string.
\item UpdateChannel — Updates a channel's configuration. This does not
affect an ongoing stream of this channel. You must stop and restart
the stream for the changes to take effect.
\item DeleteChannel — Deletes a specified channel.
}

StreamKey:
\itemize{
\item CreateStreamKey — Creates a stream key, used to initiate a stream,
for a specified channel ARN.
\item GetStreamKey — Gets stream key information for the specified ARN.
\item BatchGetStreamKey — Performs GetStreamKey on multiple ARNs
simultaneously.
\item ListStreamKeys — Gets a list of stream keys. The list can be
filtered to a particular channel.
\item DeleteStreamKey — Deletes the stream key for a specified ARN, so it
can no longer be used to stream.
}

Stream:
\itemize{
\item GetStream — Gets information about the active (live) stream on a
specified channel.
\item ListStreams — Gets summary information about live streams.
\item StopStream — Disconnects a streamer on a specified channel. This
disconnects the incoming RTMP stream from the client. Can be used in
conjunction with DeleteStreamKey to prevent further streaming to a
channel.
\item PutMetadata Inserts metadata into an RTMP stream for a specified
channel. A maximum of 5 requests per second per channel is allowed,
each with a maximum 1KB payload.
}

\href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{AWS Tags}:
\itemize{
\item TagResource — Adds or updates tags for an AWS resource with a
specified ARN.
\item UntagResource — Removes tags from a resource with a specified ARN.
\item ListTagsForResource — Gets information about AWS tags for a
specified ARN.
}
}
\section{Service syntax}{
\preformatted{svc <- ivs(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=ivs_batch_get_channel]{batch_get_channel} \tab Performs GetChannel on multiple ARNs simultaneously\cr
\link[=ivs_batch_get_stream_key]{batch_get_stream_key} \tab Performs GetStreamKey on multiple ARNs simultaneously\cr
\link[=ivs_create_channel]{create_channel} \tab Creates a new channel and an associated stream key to start streaming\cr
\link[=ivs_create_stream_key]{create_stream_key} \tab Creates a stream key, used to initiate a stream, for a specified channel ARN\cr
\link[=ivs_delete_channel]{delete_channel} \tab Deletes a specified channel and its associated stream keys\cr
\link[=ivs_delete_stream_key]{delete_stream_key} \tab Deletes the stream key for a specified ARN, so it can no longer be used to stream\cr
\link[=ivs_get_channel]{get_channel} \tab Gets the channel configuration for a specified channel ARN\cr
\link[=ivs_get_stream]{get_stream} \tab Gets information about the active (live) stream on a specified channel\cr
\link[=ivs_get_stream_key]{get_stream_key} \tab Gets stream-key information for a specified ARN\cr
\link[=ivs_list_channels]{list_channels} \tab Gets summary information about channels\cr
\link[=ivs_list_stream_keys]{list_stream_keys} \tab Gets summary information about stream keys\cr
\link[=ivs_list_streams]{list_streams} \tab Gets summary information about live streams\cr
\link[=ivs_list_tags_for_resource]{list_tags_for_resource} \tab Gets information about the tags for a specified ARN\cr
\link[=ivs_put_metadata]{put_metadata} \tab Inserts metadata into an RTMP stream for a specified channel\cr
\link[=ivs_stop_stream]{stop_stream} \tab Disconnects the stream for the specified channel\cr
\link[=ivs_tag_resource]{tag_resource} \tab Adds or updates tags for a resource with a specified ARN\cr
\link[=ivs_untag_resource]{untag_resource} \tab Removes tags for a resource with a specified ARN\cr
\link[=ivs_update_channel]{update_channel} \tab Updates a channel's configuration
}
}

\examples{
\dontrun{
svc <- ivs()
svc$batch_get_channel(
  Foo = 123
)
}

}
