% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_create_channel}
\alias{ivs_create_channel}
\title{Creates a new channel and an associated stream key to start streaming}
\usage{
ivs_create_channel(name, latencyMode, type, tags)
}
\arguments{
\item{name}{Channel name.}

\item{latencyMode}{Channel latency mode. Default: \code{LOW}.}

\item{type}{Channel type, which determines the allowable resolution and bitrate.
\code{STANDARD}: The stream is transcoded; resolution (width, in landscape
orientation) can be up to 1080p or the input source resolution,
whichever is lower; and bitrate can be up to 8.5 Mbps. \code{BASIC}: The
stream is transfixed; resolution can be up to 480p; and bitrate can be
up to 1.5 Mbps. Default: \code{STANDARD}.}

\item{tags}{See Channel$tags.}
}
\description{
Creates a new channel and an associated stream key to start streaming.
}
\section{Request syntax}{
\preformatted{svc$create_channel(
  name = "string",
  latencyMode = "NORMAL"|"LOW",
  type = "BASIC"|"STANDARD",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
