% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_tags_for_resource}
\alias{ivs_list_tags_for_resource}
\title{Gets information about the tags for a specified ARN}
\usage{
ivs_list_tags_for_resource(resourceArn, nextToken, maxResults)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be retrieved.}

\item{nextToken}{The first tag to retrieve. This is used for pagination; see the
\code{nextToken} response field.}

\item{maxResults}{Maximum number of tags to return.}
}
\description{
Gets information about the tags for a specified ARN.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
