% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_service.R
\name{kendra}
\alias{kendra}
\title{AWSKendraFrontendService}
\usage{
kendra(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Kendra is a service for indexing large document sets.
}
\section{Service syntax}{
\preformatted{svc <- kendra(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Examples}{
\preformatted{svc$batch_delete_document(
  Foo = 123
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=kendra_batch_delete_document]{batch_delete_document} \tab Removes one or more documents from an index \cr
\link[=kendra_batch_put_document]{batch_put_document} \tab Adds one or more documents to an index \cr
\link[=kendra_create_data_source]{create_data_source} \tab Creates a data source that you use to with an Amazon Kendra index \cr
\link[=kendra_create_faq]{create_faq} \tab Creates an new set of frequently asked question (FAQ) questions and answers \cr
\link[=kendra_create_index]{create_index} \tab Creates a new Amazon Kendra index \cr
\link[=kendra_delete_faq]{delete_faq} \tab Removes an FAQ from an index \cr
\link[=kendra_delete_index]{delete_index} \tab Deletes an existing Amazon Kendra index \cr
\link[=kendra_describe_data_source]{describe_data_source} \tab Gets information about a Amazon Kendra data source \cr
\link[=kendra_describe_faq]{describe_faq} \tab Gets information about an FAQ list \cr
\link[=kendra_describe_index]{describe_index} \tab Describes an existing Amazon Kendra index \cr
\link[=kendra_list_data_sources]{list_data_sources} \tab Lists the data sources that you have created \cr
\link[=kendra_list_data_source_sync_jobs]{list_data_source_sync_jobs} \tab Gets statistics about synchronizing Amazon Kendra with a data source \cr
\link[=kendra_list_faqs]{list_faqs} \tab Gets a list of FAQ lists associated with an index \cr
\link[=kendra_list_indices]{list_indices} \tab Lists the Amazon Kendra indexes that you have created \cr
\link[=kendra_query]{query} \tab Searches an active index \cr
\link[=kendra_start_data_source_sync_job]{start_data_source_sync_job} \tab Starts a synchronization job for a data source \cr
\link[=kendra_stop_data_source_sync_job]{stop_data_source_sync_job} \tab Stops a running synchronization job \cr
\link[=kendra_submit_feedback]{submit_feedback} \tab Enables you to provide feedback to Amazon Kendra to improve the performance of the service\cr
\link[=kendra_update_data_source]{update_data_source} \tab Updates an existing Amazon Kendra data source \cr
\link[=kendra_update_index]{update_index} \tab Updates an existing Amazon Kendra index
}
}

