% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_create_data_source}
\alias{kendra_create_data_source}
\title{Creates a data source that you use to with an Amazon Kendra index}
\usage{
kendra_create_data_source(Name, IndexId, Type, Configuration,
  Description, Schedule, RoleArn)
}
\arguments{
\item{Name}{[required] A unique name for the data source. A data source name can\'t be changed
without deleting and recreating the data source.}

\item{IndexId}{[required] The identifier of the index that should be associated with this data
source.}

\item{Type}{[required] The type of repository that contains the data source.}

\item{Configuration}{[required] The connector configuration information that is required to access the
repository.}

\item{Description}{A description for the data source.}

\item{Schedule}{Sets the frequency that Amazon Kendra will check the documents in your
repository and update the index. If you don\'t set a schedule Amazon
Kendra will not periodically update the index. You can call the
\code{StartDataSourceSyncJob} operation to update the index.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of a role with permission to access the
data source. For more information, see \href{https://docs.aws.amazon.com/kendra/latest/dg/iam-roles.html}{IAM Roles for Amazon Kendra}.}
}
\description{
Creates a data source that you use to with an Amazon Kendra index.
}
\details{
You specify a name, connector type and description for your data source.
You can choose between an S3 connector, a SharePoint Online connector,
and a database connector.

You also specify configuration information such as document metadata
(author, source URI, and so on) and user context information.

\code{CreateDataSource} is a synchronous operation. The operation returns 200
if the data source was successfully created. Otherwise, an exception is
raised.
}
\section{Request syntax}{
\preformatted{svc$create_data_source(
  Name = "string",
  IndexId = "string",
  Type = "S3"|"SHAREPOINT"|"DATABASE",
  Configuration = list(
    S3Configuration = list(
      BucketName = "string",
      InclusionPrefixes = list(
        "string"
      ),
      ExclusionPatterns = list(
        "string"
      ),
      DocumentsMetadataConfiguration = list(
        S3Prefix = "string"
      ),
      AccessControlListConfiguration = list(
        KeyPath = "string"
      )
    ),
    SharePointConfiguration = list(
      SharePointVersion = "SHAREPOINT_ONLINE",
      Urls = list(
        "string"
      ),
      SecretArn = "string",
      CrawlAttachments = TRUE|FALSE,
      VpcConfiguration = list(
        SubnetIds = list(
          "string"
        ),
        SecurityGroupIds = list(
          "string"
        )
      ),
      FieldMappings = list(
        list(
          DataSourceFieldName = "string",
          DateFieldFormat = "string",
          IndexFieldName = "string"
        )
      ),
      DocumentTitleFieldName = "string"
    ),
    DatabaseConfiguration = list(
      DatabaseEngineType = "RDS_AURORA_MYSQL"|"RDS_AURORA_POSTGRESQL"|"RDS_MYSQL"|"RDS_POSTGRESQL",
      ConnectionConfiguration = list(
        DatabaseHost = "string",
        DatabasePort = 123,
        DatabaseName = "string",
        TableName = "string",
        SecretArn = "string"
      ),
      VpcConfiguration = list(
        SubnetIds = list(
          "string"
        ),
        SecurityGroupIds = list(
          "string"
        )
      ),
      ColumnConfiguration = list(
        DocumentIdColumnName = "string",
        DocumentDataColumnName = "string",
        DocumentTitleColumnName = "string",
        FieldMappings = list(
          list(
            DataSourceFieldName = "string",
            DateFieldFormat = "string",
            IndexFieldName = "string"
          )
        ),
        ChangeDetectingColumns = list(
          "string"
        )
      ),
      AclConfiguration = list(
        AllowedGroupsColumnName = "string"
      )
    )
  ),
  Description = "string",
  Schedule = "string",
  RoleArn = "string"
)
}
}

\keyword{internal}
