% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_create_index}
\alias{kendra_create_index}
\title{Creates a new Amazon Kendra index}
\usage{
kendra_create_index(Name, RoleArn, ServerSideEncryptionConfiguration,
  Description)
}
\arguments{
\item{Name}{[required] The name for the new index.}

\item{RoleArn}{[required] An IAM role that gives Amazon Kendra permissions to access your Amazon
CloudWatch logs and metrics. This is also the role used when you use the
\code{BatchPutDocument} operation to index documents from an Amazon S3
bucket.}

\item{ServerSideEncryptionConfiguration}{The identifier of the KMS key to use to encrypt data indexed by Amazon
Kendra.}

\item{Description}{A description for the index.}
}
\description{
Creates a new Amazon Kendra index. Index creation is an asynchronous
operation. To determine if index creation has completed, check the
\code{Status} field returned from a call to . The \code{Status} field is set to
\code{ACTIVE} when the index is ready to use.
}
\details{
Once the index is active you can index your documents using the
operation or using one of the supported data sources.
}
\section{Request syntax}{
\preformatted{svc$create_index(
  Name = "string",
  RoleArn = "string",
  ServerSideEncryptionConfiguration = list(
    KmsKeyId = "string"
  ),
  Description = "string"
)
}
}

\keyword{internal}
