% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_create_index}
\alias{kendra_create_index}
\title{Creates a new Amazon Kendra index}
\usage{
kendra_create_index(Name, Edition, RoleArn,
  ServerSideEncryptionConfiguration, Description, ClientToken, Tags,
  UserTokenConfigurations, UserContextPolicy)
}
\arguments{
\item{Name}{[required] The name for the new index.}

\item{Edition}{The Amazon Kendra edition to use for the index. Choose
\code{DEVELOPER_EDITION} for indexes intended for development, testing, or
proof of concept. Use \code{ENTERPRISE_EDITION} for your production
databases. Once you set the edition for an index, it can't be changed.

The \code{Edition} parameter is optional. If you don't supply a value, the
default is \code{ENTERPRISE_EDITION}.}

\item{RoleArn}{[required] An AWS Identity and Access Management (IAM) role that gives Amazon
Kendra permissions to access your Amazon CloudWatch logs and metrics.
This is also the role used when you use the
\code{\link[=kendra_batch_put_document]{batch_put_document}} operation to index
documents from an Amazon S3 bucket.}

\item{ServerSideEncryptionConfiguration}{The identifier of the AWS KMS customer managed key (CMK) to use to
encrypt data indexed by Amazon Kendra. Amazon Kendra doesn't support
asymmetric CMKs.}

\item{Description}{A description for the index.}

\item{ClientToken}{A token that you provide to identify the request to create an index.
Multiple calls to the \code{\link[=kendra_create_index]{create_index}} operation
with the same client token will create only one index.}

\item{Tags}{A list of key-value pairs that identify the index. You can use the tags
to identify and organize your resources and to control access to
resources.}

\item{UserTokenConfigurations}{The user token configuration.}

\item{UserContextPolicy}{The user context policy.
\subsection{ATTRIBUTE\\_FILTER}{

All indexed content is searchable and displayable for all users. If
there is an access control list, it is ignored. You can filter on user
and group attributes.
}

\subsection{USER\\_TOKEN}{

Enables SSO and token-based user access control. All documents with no
access control and all documents accessible to the user will be
searchable and displayable.
}}
}
\description{
Creates a new Amazon Kendra index. Index creation is an asynchronous
operation. To determine if index creation has completed, check the
\code{Status} field returned from a call to . The \code{Status} field is set to
\code{ACTIVE} when the index is ready to use.

Once the index is active you can index your documents using the
operation or using one of the supported data sources.
}
\section{Request syntax}{
\preformatted{svc$create_index(
  Name = "string",
  Edition = "DEVELOPER_EDITION"|"ENTERPRISE_EDITION",
  RoleArn = "string",
  ServerSideEncryptionConfiguration = list(
    KmsKeyId = "string"
  ),
  Description = "string",
  ClientToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  UserTokenConfigurations = list(
    list(
      JwtTokenTypeConfiguration = list(
        KeyLocation = "URL"|"SECRET_MANAGER",
        URL = "string",
        SecretManagerArn = "string",
        UserNameAttributeField = "string",
        GroupAttributeField = "string",
        Issuer = "string",
        ClaimRegex = "string"
      ),
      JsonTokenTypeConfiguration = list(
        UserNameAttributeField = "string",
        GroupAttributeField = "string"
      )
    )
  ),
  UserContextPolicy = "ATTRIBUTE_FILTER"|"USER_TOKEN"
)
}
}

\keyword{internal}
