% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_update_data_source}
\alias{kendra_update_data_source}
\title{Updates an existing Amazon Kendra data source}
\usage{
kendra_update_data_source(Id, Name, IndexId, Configuration, Description,
  Schedule, RoleArn)
}
\arguments{
\item{Id}{[required] The unique identifier of the data source to update.}

\item{Name}{The name of the data source to update. The name of the data source
can\'t be updated. To rename a data source you must delete the data
source and re-create it.}

\item{IndexId}{[required] The identifier of the index that contains the data source to update.}

\item{Configuration}{}

\item{Description}{The new description for the data source.}

\item{Schedule}{The new update schedule for the data source.}

\item{RoleArn}{The Amazon Resource Name (ARN) of the new role to use when the data
source is accessing resources on your behalf.}
}
\description{
Updates an existing Amazon Kendra data source.
}
\section{Request syntax}{
\preformatted{svc$update_data_source(
  Id = "string",
  Name = "string",
  IndexId = "string",
  Configuration = list(
    S3Configuration = list(
      BucketName = "string",
      InclusionPrefixes = list(
        "string"
      ),
      ExclusionPatterns = list(
        "string"
      ),
      DocumentsMetadataConfiguration = list(
        S3Prefix = "string"
      ),
      AccessControlListConfiguration = list(
        KeyPath = "string"
      )
    ),
    SharePointConfiguration = list(
      SharePointVersion = "SHAREPOINT_ONLINE",
      Urls = list(
        "string"
      ),
      SecretArn = "string",
      CrawlAttachments = TRUE|FALSE,
      UseChangeLog = TRUE|FALSE,
      InclusionPatterns = list(
        "string"
      ),
      ExclusionPatterns = list(
        "string"
      ),
      VpcConfiguration = list(
        SubnetIds = list(
          "string"
        ),
        SecurityGroupIds = list(
          "string"
        )
      ),
      FieldMappings = list(
        list(
          DataSourceFieldName = "string",
          DateFieldFormat = "string",
          IndexFieldName = "string"
        )
      ),
      DocumentTitleFieldName = "string"
    ),
    DatabaseConfiguration = list(
      DatabaseEngineType = "RDS_AURORA_MYSQL"|"RDS_AURORA_POSTGRESQL"|"RDS_MYSQL"|"RDS_POSTGRESQL",
      ConnectionConfiguration = list(
        DatabaseHost = "string",
        DatabasePort = 123,
        DatabaseName = "string",
        TableName = "string",
        SecretArn = "string"
      ),
      VpcConfiguration = list(
        SubnetIds = list(
          "string"
        ),
        SecurityGroupIds = list(
          "string"
        )
      ),
      ColumnConfiguration = list(
        DocumentIdColumnName = "string",
        DocumentDataColumnName = "string",
        DocumentTitleColumnName = "string",
        FieldMappings = list(
          list(
            DataSourceFieldName = "string",
            DateFieldFormat = "string",
            IndexFieldName = "string"
          )
        ),
        ChangeDetectingColumns = list(
          "string"
        )
      ),
      AclConfiguration = list(
        AllowedGroupsColumnName = "string"
      )
    ),
    SalesforceConfiguration = list(
      ServerUrl = "string",
      SecretArn = "string",
      StandardObjectConfigurations = list(
        list(
          Name = "ACCOUNT"|"CAMPAIGN"|"CASE"|"CONTACT"|"CONTRACT"|"DOCUMENT"|"GROUP"|"IDEA"|"LEAD"|"OPPORTUNITY"|"PARTNER"|"PRICEBOOK"|"PRODUCT"|"PROFILE"|"SOLUTION"|"TASK"|"USER",
          DocumentDataFieldName = "string",
          DocumentTitleFieldName = "string",
          FieldMappings = list(
            list(
              DataSourceFieldName = "string",
              DateFieldFormat = "string",
              IndexFieldName = "string"
            )
          )
        )
      ),
      KnowledgeArticleConfiguration = list(
        IncludedStates = list(
          "DRAFT"|"PUBLISHED"|"ARCHIVED"
        ),
        StandardKnowledgeArticleTypeConfiguration = list(
          DocumentDataFieldName = "string",
          DocumentTitleFieldName = "string",
          FieldMappings = list(
            list(
              DataSourceFieldName = "string",
              DateFieldFormat = "string",
              IndexFieldName = "string"
            )
          )
        ),
        CustomKnowledgeArticleTypeConfigurations = list(
          list(
            Name = "string",
            DocumentDataFieldName = "string",
            DocumentTitleFieldName = "string",
            FieldMappings = list(
              list(
                DataSourceFieldName = "string",
                DateFieldFormat = "string",
                IndexFieldName = "string"
              )
            )
          )
        )
      ),
      ChatterFeedConfiguration = list(
        DocumentDataFieldName = "string",
        DocumentTitleFieldName = "string",
        FieldMappings = list(
          list(
            DataSourceFieldName = "string",
            DateFieldFormat = "string",
            IndexFieldName = "string"
          )
        ),
        IncludeFilterTypes = list(
          "ACTIVE_USER"|"STANDARD_USER"
        )
      ),
      CrawlAttachments = TRUE|FALSE,
      StandardObjectAttachmentConfiguration = list(
        DocumentTitleFieldName = "string",
        FieldMappings = list(
          list(
            DataSourceFieldName = "string",
            DateFieldFormat = "string",
            IndexFieldName = "string"
          )
        )
      ),
      IncludeAttachmentFilePatterns = list(
        "string"
      ),
      ExcludeAttachmentFilePatterns = list(
        "string"
      )
    ),
    OneDriveConfiguration = list(
      TenantDomain = "string",
      SecretArn = "string",
      OneDriveUsers = list(
        OneDriveUserList = list(
          "string"
        ),
        OneDriveUserS3Path = list(
          Bucket = "string",
          Key = "string"
        )
      ),
      InclusionPatterns = list(
        "string"
      ),
      ExclusionPatterns = list(
        "string"
      ),
      FieldMappings = list(
        list(
          DataSourceFieldName = "string",
          DateFieldFormat = "string",
          IndexFieldName = "string"
        )
      )
    ),
    ServiceNowConfiguration = list(
      HostUrl = "string",
      SecretArn = "string",
      ServiceNowBuildVersion = "LONDON"|"OTHERS",
      KnowledgeArticleConfiguration = list(
        CrawlAttachments = TRUE|FALSE,
        IncludeAttachmentFilePatterns = list(
          "string"
        ),
        ExcludeAttachmentFilePatterns = list(
          "string"
        ),
        DocumentDataFieldName = "string",
        DocumentTitleFieldName = "string",
        FieldMappings = list(
          list(
            DataSourceFieldName = "string",
            DateFieldFormat = "string",
            IndexFieldName = "string"
          )
        )
      ),
      ServiceCatalogConfiguration = list(
        CrawlAttachments = TRUE|FALSE,
        IncludeAttachmentFilePatterns = list(
          "string"
        ),
        ExcludeAttachmentFilePatterns = list(
          "string"
        ),
        DocumentDataFieldName = "string",
        DocumentTitleFieldName = "string",
        FieldMappings = list(
          list(
            DataSourceFieldName = "string",
            DateFieldFormat = "string",
            IndexFieldName = "string"
          )
        )
      )
    )
  ),
  Description = "string",
  Schedule = "string",
  RoleArn = "string"
)
}
}

\keyword{internal}
