% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{kinesis}
\alias{kinesis}
\title{Amazon Kinesis}
\usage{
kinesis(config = list(), credentials = list(), endpoint = NULL, region = NULL)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Amazon Kinesis Data Streams Service API Reference

Amazon Kinesis Data Streams is a managed service that scales elastically for real-time processing of streaming big data.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- kinesis(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.analytics:kinesis_add_tags_to_stream]{add_tags_to_stream} \tab Adds or updates tags for the specified Kinesis data stream\cr
\link[paws.analytics:kinesis_create_channel]{create_channel} \tab Creates a channel that delivers records from a Kinesis data stream to a destination\cr
\link[paws.analytics:kinesis_create_stream]{create_stream} \tab Creates a Kinesis data stream\cr
\link[paws.analytics:kinesis_decrease_stream_retention_period]{decrease_stream_retention_period} \tab Decreases the Kinesis data stream's retention period, which is the length of time data records are accessible after they are added to the stream\cr
\link[paws.analytics:kinesis_delete_channel]{delete_channel} \tab Deletes the specified channel\cr
\link[paws.analytics:kinesis_delete_resource_policy]{delete_resource_policy} \tab Delete a policy for the specified data stream or consumer\cr
\link[paws.analytics:kinesis_delete_stream]{delete_stream} \tab Deletes a Kinesis data stream and all its shards and data\cr
\link[paws.analytics:kinesis_deregister_stream_consumer]{deregister_stream_consumer} \tab To deregister a consumer, provide its ARN\cr
\link[paws.analytics:kinesis_describe_account_settings]{describe_account_settings} \tab Describes the account-level settings for Amazon Kinesis Data Streams\cr
\link[paws.analytics:kinesis_describe_channel]{describe_channel} \tab Describes the specified channel, including its configuration and current status\cr
\link[paws.analytics:kinesis_describe_limits]{describe_limits} \tab Describes the shard limits and usage for the account\cr
\link[paws.analytics:kinesis_describe_stream]{describe_stream} \tab Describes the specified Kinesis data stream\cr
\link[paws.analytics:kinesis_describe_stream_consumer]{describe_stream_consumer} \tab To get the description of a registered consumer, provide the ARN of the consumer\cr
\link[paws.analytics:kinesis_describe_stream_summary]{describe_stream_summary} \tab Provides a summarized description of the specified Kinesis data stream without the shard list\cr
\link[paws.analytics:kinesis_disable_enhanced_monitoring]{disable_enhanced_monitoring} \tab Disables enhanced monitoring\cr
\link[paws.analytics:kinesis_enable_enhanced_monitoring]{enable_enhanced_monitoring} \tab Enables enhanced Kinesis data stream monitoring for shard-level metrics\cr
\link[paws.analytics:kinesis_get_records]{get_records} \tab Gets data records from a Kinesis data stream's shard\cr
\link[paws.analytics:kinesis_get_resource_policy]{get_resource_policy} \tab Returns a policy attached to the specified data stream or consumer\cr
\link[paws.analytics:kinesis_get_shard_iterator]{get_shard_iterator} \tab Gets an Amazon Kinesis shard iterator\cr
\link[paws.analytics:kinesis_increase_stream_retention_period]{increase_stream_retention_period} \tab Increases the Kinesis data stream's retention period, which is the length of time data records are accessible after they are added to the stream\cr
\link[paws.analytics:kinesis_list_channels]{list_channels} \tab Lists the channels in your account\cr
\link[paws.analytics:kinesis_list_shards]{list_shards} \tab Lists the shards in a stream and provides information about each shard\cr
\link[paws.analytics:kinesis_list_stream_consumers]{list_stream_consumers} \tab Lists the consumers registered to receive data from a stream using enhanced fan-out, and provides information about each consumer\cr
\link[paws.analytics:kinesis_list_streams]{list_streams} \tab Lists your Kinesis data streams\cr
\link[paws.analytics:kinesis_list_tags_for_resource]{list_tags_for_resource} \tab List all tags added to the specified Kinesis resource\cr
\link[paws.analytics:kinesis_list_tags_for_stream]{list_tags_for_stream} \tab Lists the tags for the specified Kinesis data stream\cr
\link[paws.analytics:kinesis_merge_shards]{merge_shards} \tab Merges two adjacent shards in a Kinesis data stream and combines them into a single shard to reduce the stream's capacity to ingest and transport data\cr
\link[paws.analytics:kinesis_put_record]{put_record} \tab Writes a single data record into an Amazon Kinesis data stream\cr
\link[paws.analytics:kinesis_put_records]{put_records} \tab Writes multiple data records into a Kinesis data stream in a single call (also referred to as a PutRecords request)\cr
\link[paws.analytics:kinesis_put_resource_policy]{put_resource_policy} \tab Attaches a resource-based policy to a data stream or registered consumer\cr
\link[paws.analytics:kinesis_register_stream_consumer]{register_stream_consumer} \tab Registers a consumer with a Kinesis data stream\cr
\link[paws.analytics:kinesis_remove_tags_from_stream]{remove_tags_from_stream} \tab Removes tags from the specified Kinesis data stream\cr
\link[paws.analytics:kinesis_split_shard]{split_shard} \tab Splits a shard into two new shards in the Kinesis data stream, to increase the stream's capacity to ingest and transport data\cr
\link[paws.analytics:kinesis_start_stream_encryption]{start_stream_encryption} \tab Enables or updates server-side encryption using an Amazon Web Services KMS key for a specified stream\cr
\link[paws.analytics:kinesis_stop_stream_encryption]{stop_stream_encryption} \tab Disables server-side encryption for a specified stream\cr
\link[paws.analytics:kinesis_subscribe_to_shard]{subscribe_to_shard} \tab This operation establishes an HTTP/2 connection between the consumer you specify in the ConsumerARN parameter and the shard you specify in the ShardId parameter\cr
\link[paws.analytics:kinesis_tag_resource]{tag_resource} \tab Adds or updates tags for the specified Kinesis resource\cr
\link[paws.analytics:kinesis_untag_resource]{untag_resource} \tab Removes tags from the specified Kinesis resource\cr
\link[paws.analytics:kinesis_update_account_settings]{update_account_settings} \tab Updates the account-level settings for Amazon Kinesis Data Streams\cr
\link[paws.analytics:kinesis_update_channel]{update_channel} \tab Updates the data freshness interval or the Amazon CloudWatch Logs configuration of an existing channel\cr
\link[paws.analytics:kinesis_update_max_record_size]{update_max_record_size} \tab This allows you to update the MaxRecordSize of a single record that you can write to, and read from a stream\cr
\link[paws.analytics:kinesis_update_shard_count]{update_shard_count} \tab Updates the shard count of the specified stream to the specified number of shards\cr
\link[paws.analytics:kinesis_update_stream_mode]{update_stream_mode} \tab Updates the capacity mode of the data stream\cr
\link[paws.analytics:kinesis_update_stream_warm_throughput]{update_stream_warm_throughput} \tab Updates the warm throughput configuration for the specified Amazon Kinesis Data Streams on-demand data stream
}
}

\examples{
\dontrun{
svc <- kinesis()
svc$add_tags_to_stream(
  Foo = 123
)
}

}
