% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_disable_key}
\alias{kms_disable_key}
\title{Sets the state of a customer master key (CMK) to disabled, thereby
preventing its use for cryptographic operations}
\usage{
kms_disable_key(KeyId)
}
\arguments{
\item{KeyId}{[required] A unique identifier for the customer master key (CMK).

Specify the key ID or the Amazon Resource Name (ARN) of the CMK.

For example:
\itemize{
\item Key ID: \code{1234abcd-12ab-34cd-56ef-1234567890ab}
\item Key ARN:
\code{arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab}
}

To get the key ID and key ARN for a CMK, use ListKeys or DescribeKey.}
}
\description{
Sets the state of a customer master key (CMK) to disabled, thereby
preventing its use for cryptographic operations. You cannot perform this
operation on a CMK in a different AWS account.
}
\details{
For more information about how key state affects the use of a CMK, see
\href{https://docs.aws.amazon.com/kms/latest/developerguide/key-state.html}{How Key State Affects the Use of a Customer Master Key}
in the \strong{AWS Key Management Service Developer Guide} .

The result of this operation varies with the key state of the CMK. For
details, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/key-state.html}{How Key State Affects Use of a Customer Master Key}
in the \emph{AWS Key Management Service Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$disable_key(
  KeyId = "string"
)
}
}

\examples{
# The following example disables the specified CMK.
\donttest{svc$disable_key(
  KeyId = "1234abcd-12ab-34cd-56ef-1234567890ab"
)}

}
\keyword{internal}
