% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_service.R
\name{lakeformation}
\alias{lakeformation}
\title{AWS Lake Formation}
\usage{
lakeformation(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Defines the public endpoint for the AWS Lake Formation service.
}
\section{Service syntax}{
\preformatted{svc <- lakeformation(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=lakeformation_batch_grant_permissions]{batch_grant_permissions} \tab Batch operation to grant permissions to the principal\cr
\link[=lakeformation_batch_revoke_permissions]{batch_revoke_permissions} \tab Batch operation to revoke permissions from the principal\cr
\link[=lakeformation_deregister_resource]{deregister_resource} \tab Deregisters the resource as managed by the Data Catalog\cr
\link[=lakeformation_describe_resource]{describe_resource} \tab Retrieves the current data access role for the given resource registered in AWS Lake Formation\cr
\link[=lakeformation_get_data_lake_settings]{get_data_lake_settings} \tab Retrieves the list of the data lake administrators of a Lake Formation-managed data lake\cr
\link[=lakeformation_get_effective_permissions_for_path]{get_effective_permissions_for_path} \tab Returns the Lake Formation permissions for a specified table or database resource located at a path in Amazon S3\cr
\link[=lakeformation_grant_permissions]{grant_permissions} \tab Grants permissions to the principal to access metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3\cr
\link[=lakeformation_list_permissions]{list_permissions} \tab Returns a list of the principal permissions on the resource, filtered by the permissions of the caller\cr
\link[=lakeformation_list_resources]{list_resources} \tab Lists the resources registered to be managed by the Data Catalog\cr
\link[=lakeformation_put_data_lake_settings]{put_data_lake_settings} \tab Sets the list of data lake administrators who have admin privileges on all resources managed by Lake Formation\cr
\link[=lakeformation_register_resource]{register_resource} \tab Registers the resource as managed by the Data Catalog\cr
\link[=lakeformation_revoke_permissions]{revoke_permissions} \tab Revokes permissions to the principal to access metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3\cr
\link[=lakeformation_update_resource]{update_resource} \tab Updates the data access role used for vending access to the given (registered) resource in AWS Lake Formation
}
}

\examples{
\dontrun{
svc <- lakeformation()
svc$batch_grant_permissions(
  Foo = 123
)
}

}
