% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_license_configuration}
\alias{licensemanager_get_license_configuration}
\title{Returns a detailed description of a license configuration}
\usage{
licensemanager_get_license_configuration(LicenseConfigurationArn)
}
\arguments{
\item{LicenseConfigurationArn}{[required] ARN of the license configuration being requested.}
}
\description{
Returns a detailed description of a license configuration.
}
\section{Request syntax}{
\preformatted{licensemanager$get_license_configuration(
  LicenseConfigurationArn = "string"
)
}
}

\keyword{internal}
