% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_service_settings}
\alias{licensemanager_get_service_settings}
\title{Gets the License Manager settings for the current Region}
\usage{
licensemanager_get_service_settings()
}
\description{
Gets the License Manager settings for the current Region.
}
\section{Request syntax}{
\preformatted{svc$get_service_settings()
}
}

\keyword{internal}
