% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_operations}
\alias{lightsail_get_operations}
\title{Returns information about all operations}
\usage{
lightsail_get_operations(pageToken)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial \code{GetOperations} request. If your
results are paginated, the response will return a next page token that
you can specify as the page token in a subsequent request.}
}
\description{
Returns information about all operations.
}
\details{
Results are returned from oldest to newest, up to a maximum of 200.
Results can be paged by making each subsequent call to \code{GetOperations}
use the maximum (last) \code{statusChangedAt} value from the previous
request.
}
\section{Request syntax}{
\preformatted{svc$get_operations(
  pageToken = "string"
)
}
}

\keyword{internal}
