% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database_snapshots}
\alias{lightsail_get_relational_database_snapshots}
\title{Returns information about all of your database snapshots in Amazon
Lightsail}
\usage{
lightsail_get_relational_database_snapshots(pageToken)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_relational_database_snapshots]{get_relational_database_snapshots}}
request. If your results are paginated, the response will return a next
page token that you can specify as the page token in a subsequent
request.}
}
\description{
Returns information about all of your database snapshots in Amazon
Lightsail.
}
\section{Request syntax}{
\preformatted{svc$get_relational_database_snapshots(
  pageToken = "string"
)
}
}

\keyword{internal}
