% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_service.R
\name{locationservice}
\alias{locationservice}
\title{Amazon Location Service}
\usage{
locationservice(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Suite of geospatial services including Maps, Places, Tracking, and
Geofencing
}
\section{Service syntax}{
\preformatted{svc <- locationservice(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=locationservice_associate_tracker_consumer]{associate_tracker_consumer} \tab Creates an association between a geofence collection and a tracker resource\cr
\link[=locationservice_batch_delete_geofence]{batch_delete_geofence} \tab Deletes a batch of geofences from a geofence collection\cr
\link[=locationservice_batch_evaluate_geofences]{batch_evaluate_geofences} \tab Used in geofence monitoring\cr
\link[=locationservice_batch_get_device_position]{batch_get_device_position} \tab A batch request to retrieve device positions\cr
\link[=locationservice_batch_put_geofence]{batch_put_geofence} \tab A batch request for storing geofences into a given geofence collection\cr
\link[=locationservice_batch_update_device_position]{batch_update_device_position} \tab Uploads a position update for one or more devices to a tracker resource\cr
\link[=locationservice_create_geofence_collection]{create_geofence_collection} \tab Creates a geofence collection, which manages and stores geofences\cr
\link[=locationservice_create_map]{create_map} \tab Creates a map resource in your AWS account, which provides map tiles of different styles sourced from global location data providers\cr
\link[=locationservice_create_place_index]{create_place_index} \tab Creates a Place index resource in your AWS account, which supports Places functions with geospatial data sourced from your chosen data provider\cr
\link[=locationservice_create_tracker]{create_tracker} \tab Creates a tracker resource in your AWS account, which lets you retrieve current and historical location of devices\cr
\link[=locationservice_delete_geofence_collection]{delete_geofence_collection} \tab Deletes a geofence collection from your AWS account\cr
\link[=locationservice_delete_map]{delete_map} \tab Deletes a map resource from your AWS account\cr
\link[=locationservice_delete_place_index]{delete_place_index} \tab Deletes a Place index resource from your AWS account\cr
\link[=locationservice_delete_tracker]{delete_tracker} \tab Deletes a tracker resource from your AWS account\cr
\link[=locationservice_describe_geofence_collection]{describe_geofence_collection} \tab Retrieves the geofence collection details\cr
\link[=locationservice_describe_map]{describe_map} \tab Retrieves the map resource details\cr
\link[=locationservice_describe_place_index]{describe_place_index} \tab Retrieves the Place index resource details\cr
\link[=locationservice_describe_tracker]{describe_tracker} \tab Retrieves the tracker resource details\cr
\link[=locationservice_disassociate_tracker_consumer]{disassociate_tracker_consumer} \tab Removes the association bewteen a tracker resource and a geofence collection\cr
\link[=locationservice_get_device_position]{get_device_position} \tab Retrieves the latest device position\cr
\link[=locationservice_get_device_position_history]{get_device_position_history} \tab Retrieves the device position history from a tracker resource within a specified range of time\cr
\link[=locationservice_get_geofence]{get_geofence} \tab Retrieves the geofence details from a geofence collection\cr
\link[=locationservice_get_map_glyphs]{get_map_glyphs} \tab Retrieves glyphs used to display labels on a map\cr
\link[=locationservice_get_map_sprites]{get_map_sprites} \tab Retrieves the sprite sheet corresponding to a map resource\cr
\link[=locationservice_get_map_style_descriptor]{get_map_style_descriptor} \tab Retrieves the map style descriptor from a map resource\cr
\link[=locationservice_get_map_tile]{get_map_tile} \tab Retrieves a vector data tile from the map resource\cr
\link[=locationservice_list_geofence_collections]{list_geofence_collections} \tab Lists geofence collections in your AWS account\cr
\link[=locationservice_list_geofences]{list_geofences} \tab Lists geofences stored in a given geofence collection\cr
\link[=locationservice_list_maps]{list_maps} \tab Lists map resources in your AWS account\cr
\link[=locationservice_list_place_indexes]{list_place_indexes} \tab Lists Place index resources in your AWS account\cr
\link[=locationservice_list_tracker_consumers]{list_tracker_consumers} \tab Lists geofence collections currently associated to the given tracker resource\cr
\link[=locationservice_list_trackers]{list_trackers} \tab Lists tracker resources in your AWS account\cr
\link[=locationservice_put_geofence]{put_geofence} \tab Stores a geofence to a given geofence collection, or updates the geometry of an existing geofence if a geofence ID is included in the request\cr
\link[=locationservice_search_place_index_for_position]{search_place_index_for_position} \tab Reverse geocodes a given coordinate and returns a legible address\cr
\link[=locationservice_search_place_index_for_text]{search_place_index_for_text} \tab Geocodes free-form text, such as an address, name, city, or region to allow you to search for Places or points of interest
}
}

\examples{
\dontrun{
svc <- locationservice()
svc$associate_tracker_consumer(
  Foo = 123
)
}

}
