% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_get_node}
\alias{managedblockchain_get_node}
\title{Returns detailed information about a node}
\usage{
managedblockchain_get_node(NetworkId, MemberId, NodeId)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network that the node is on.}

\item{MemberId}{The unique identifier of the member that owns the node.

Applies only to Hyperledger Fabric and is required for Hyperledger
Fabric.}

\item{NodeId}{[required] The unique identifier of the node.}
}
\description{
Returns detailed information about a node.
}
\details{
Applies to Hyperledger Fabric and Ethereum.
}
\section{Request syntax}{
\preformatted{svc$get_node(
  NetworkId = "string",
  MemberId = "string",
  NodeId = "string"
)
}
}

\keyword{internal}
