% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_list_invitations}
\alias{managedblockchain_list_invitations}
\title{Returns a listing of all invitations made on the specified network}
\usage{
managedblockchain_list_invitations(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of invitations to return.}

\item{NextToken}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Returns a listing of all invitations made on the specified network.
}
\section{Request syntax}{
\preformatted{svc$list_invitations(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
