% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_list_nodes}
\alias{managedblockchain_list_nodes}
\title{Returns information about the nodes within a network}
\usage{
managedblockchain_list_nodes(NetworkId, MemberId, Status, MaxResults,
  NextToken)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network for which to list nodes.}

\item{MemberId}{[required] The unique identifier of the member who owns the nodes to list.}

\item{Status}{An optional status specifier. If provided, only nodes currently in this
status are listed.}

\item{MaxResults}{The maximum number of nodes to list.}

\item{NextToken}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Returns information about the nodes within a network.
}
\section{Request syntax}{
\preformatted{svc$list_nodes(
  NetworkId = "string",
  MemberId = "string",
  Status = "CREATING"|"AVAILABLE"|"CREATE_FAILED"|"DELETING"|"DELETED"|"FAILED",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
