% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_update_node}
\alias{managedblockchain_update_node}
\title{Updates a node configuration with new parameters}
\usage{
managedblockchain_update_node(NetworkId, MemberId, NodeId,
  LogPublishingConfiguration)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network that the node is on.}

\item{MemberId}{The unique identifier of the member that owns the node.

Applies only to Hyperledger Fabric.}

\item{NodeId}{[required] The unique identifier of the node.}

\item{LogPublishingConfiguration}{Configuration properties for publishing to Amazon CloudWatch Logs.}
}
\description{
Updates a node configuration with new parameters.
}
\details{
Applies only to Hyperledger Fabric.
}
\section{Request syntax}{
\preformatted{svc$update_node(
  NetworkId = "string",
  MemberId = "string",
  NodeId = "string",
  LogPublishingConfiguration = list(
    Fabric = list(
      ChaincodeLogs = list(
        Cloudwatch = list(
          Enabled = TRUE|FALSE
        )
      ),
      PeerLogs = list(
        Cloudwatch = list(
          Enabled = TRUE|FALSE
        )
      )
    )
  )
)
}
}

\keyword{internal}
