% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_service.R
\name{marketplacecatalog}
\alias{marketplacecatalog}
\title{AWS Marketplace Catalog Service}
\usage{
marketplacecatalog(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Catalog API actions allow you to manage your entities through list,
describe, and update capabilities. An entity can be a product or an
offer on AWS Marketplace.

You can automate your entity update process by integrating the AWS
Marketplace Catalog API with your AWS Marketplace product build or
deployment pipelines. You can also create your own applications on top
of the Catalog API to manage your products on AWS Marketplace.
}
\section{Service syntax}{
\preformatted{svc <- marketplacecatalog(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=marketplacecatalog_cancel_change_set]{cancel_change_set} \tab Used to cancel an open change request \cr
\link[=marketplacecatalog_describe_change_set]{describe_change_set} \tab Provides information about a given change set \cr
\link[=marketplacecatalog_describe_entity]{describe_entity} \tab Returns the metadata and content of the entity \cr
\link[=marketplacecatalog_list_change_sets]{list_change_sets} \tab Returns the list of change sets owned by the account being used to make the call\cr
\link[=marketplacecatalog_list_entities]{list_entities} \tab Provides the list of entities of a given type \cr
\link[=marketplacecatalog_start_change_set]{start_change_set} \tab This operation allows you to request changes for your entities
}
}

\examples{
\dontrun{
svc <- marketplacecatalog()
svc$cancel_change_set(
  Foo = 123
)
}

}
