% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackagevod_operations.R
\name{mediapackagevod_create_packaging_group}
\alias{mediapackagevod_create_packaging_group}
\title{Creates a new MediaPackage VOD PackagingGroup resource}
\usage{
mediapackagevod_create_packaging_group(Id)
}
\arguments{
\item{Id}{[required] The ID of the PackagingGroup.}
}
\description{
Creates a new MediaPackage VOD PackagingGroup resource.
}
\section{Request syntax}{
\preformatted{svc$create_packaging_group(
  Id = "string"
)
}
}

\keyword{internal}
