% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_service.R
\name{mwaa}
\alias{mwaa}
\title{AmazonMWAA}
\usage{
mwaa(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Managed Workflows for Apache Airflow

This section contains the Amazon Managed Workflows for Apache Airflow
(MWAA) API reference documentation. For more information, see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/what-is-mwaa.html}{What Is Amazon MWAA?}.
}
\section{Service syntax}{
\preformatted{svc <- mwaa(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=mwaa_create_cli_token]{create_cli_token} \tab Create a CLI token to use Airflow CLI\cr
\link[=mwaa_create_environment]{create_environment} \tab JSON blob that describes the environment to create\cr
\link[=mwaa_create_web_login_token]{create_web_login_token} \tab Create a JWT token to be used to login to Airflow Web UI with claims based Authentication\cr
\link[=mwaa_delete_environment]{delete_environment} \tab Delete an existing environment\cr
\link[=mwaa_get_environment]{get_environment} \tab Get details of an existing environment\cr
\link[=mwaa_list_environments]{list_environments} \tab List Amazon MWAA Environments\cr
\link[=mwaa_list_tags_for_resource]{list_tags_for_resource} \tab List the tags for MWAA environments\cr
\link[=mwaa_publish_metrics]{publish_metrics} \tab An operation for publishing metrics from the customers to the Ops plane\cr
\link[=mwaa_tag_resource]{tag_resource} \tab Add tag to the MWAA environments\cr
\link[=mwaa_untag_resource]{untag_resource} \tab Remove a tag from the MWAA environments\cr
\link[=mwaa_update_environment]{update_environment} \tab Update an MWAA environment
}
}

\examples{
\dontrun{
svc <- mwaa()
svc$create_cli_token(
  Foo = 123
)
}

}
