% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_service.R
\name{networkmanager}
\alias{networkmanager}
\title{AWS Network Manager}
\usage{
networkmanager(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Transit Gateway Network Manager (Network Manager) enables you to create
a global network, in which you can monitor your AWS and on-premises
networks that are built around transit gateways.
}
\section{Service syntax}{
\preformatted{svc <- networkmanager(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Examples}{
\preformatted{svc$associate_customer_gateway(
  Foo = 123
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=networkmanager_associate_customer_gateway]{associate_customer_gateway} \tab Associates a customer gateway with a device and optionally, with a link \cr
\link[=networkmanager_associate_link]{associate_link} \tab Associates a link to a device \cr
\link[=networkmanager_create_device]{create_device} \tab Creates a new device in a global network \cr
\link[=networkmanager_create_global_network]{create_global_network} \tab Creates a new, empty global network \cr
\link[=networkmanager_create_link]{create_link} \tab Creates a new link for a specified site \cr
\link[=networkmanager_create_site]{create_site} \tab Creates a new site in a global network \cr
\link[=networkmanager_delete_device]{delete_device} \tab Deletes an existing device \cr
\link[=networkmanager_delete_global_network]{delete_global_network} \tab Deletes an existing global network \cr
\link[=networkmanager_delete_link]{delete_link} \tab Deletes an existing link \cr
\link[=networkmanager_delete_site]{delete_site} \tab Deletes an existing site \cr
\link[=networkmanager_deregister_transit_gateway]{deregister_transit_gateway} \tab Deregisters a transit gateway from your global network \cr
\link[=networkmanager_describe_global_networks]{describe_global_networks} \tab Describes one or more global networks \cr
\link[=networkmanager_disassociate_customer_gateway]{disassociate_customer_gateway} \tab Disassociates a customer gateway from a device and a link \cr
\link[=networkmanager_disassociate_link]{disassociate_link} \tab Disassociates an existing device from a link \cr
\link[=networkmanager_get_customer_gateway_associations]{get_customer_gateway_associations} \tab Gets the association information for customer gateways that are associated with devices and links in your global network\cr
\link[=networkmanager_get_devices]{get_devices} \tab Gets information about one or more of your devices in a global network \cr
\link[=networkmanager_get_link_associations]{get_link_associations} \tab Gets the link associations for a device or a link \cr
\link[=networkmanager_get_links]{get_links} \tab Gets information about one or more links in a specified global network \cr
\link[=networkmanager_get_sites]{get_sites} \tab Gets information about one or more of your sites in a global network \cr
\link[=networkmanager_get_transit_gateway_registrations]{get_transit_gateway_registrations} \tab Gets information about the transit gateway registrations in a specified global network \cr
\link[=networkmanager_list_tags_for_resource]{list_tags_for_resource} \tab Lists the tags for a specified resource \cr
\link[=networkmanager_register_transit_gateway]{register_transit_gateway} \tab Registers a transit gateway in your global network \cr
\link[=networkmanager_tag_resource]{tag_resource} \tab Tags a specified resource \cr
\link[=networkmanager_untag_resource]{untag_resource} \tab Removes tags from a specified resource \cr
\link[=networkmanager_update_device]{update_device} \tab Updates the details for an existing device \cr
\link[=networkmanager_update_global_network]{update_global_network} \tab Updates an existing global network \cr
\link[=networkmanager_update_link]{update_link} \tab Updates the details for an existing link \cr
\link[=networkmanager_update_site]{update_site} \tab Updates the information for an existing site
}
}

