% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{omics}
\alias{omics}
\title{Amazon Omics}
\usage{
omics(config = list(), credentials = list(), endpoint = NULL, region = NULL)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
This is the \emph{AWS HealthOmics API Reference}. For an introduction to the
service, see \href{https://docs.aws.amazon.com/omics/latest/dev/}{What is AWS HealthOmics?} in the \emph{AWS
HealthOmics User Guide}.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- omics(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[paws.storage:omics_abort_multipart_read_set_upload]{abort_multipart_read_set_upload} \tab Stops a multipart upload\cr
\link[paws.storage:omics_accept_share]{accept_share} \tab Accepts a share for an analytics store\cr
\link[paws.storage:omics_batch_delete_read_set]{batch_delete_read_set} \tab Deletes one or more read sets\cr
\link[paws.storage:omics_cancel_annotation_import_job]{cancel_annotation_import_job} \tab Cancels an annotation import job\cr
\link[paws.storage:omics_cancel_run]{cancel_run} \tab Cancels a run\cr
\link[paws.storage:omics_cancel_variant_import_job]{cancel_variant_import_job} \tab Cancels a variant import job\cr
\link[paws.storage:omics_complete_multipart_read_set_upload]{complete_multipart_read_set_upload} \tab Concludes a multipart upload once you have uploaded all the components\cr
\link[paws.storage:omics_create_annotation_store]{create_annotation_store} \tab Creates an annotation store\cr
\link[paws.storage:omics_create_annotation_store_version]{create_annotation_store_version} \tab Creates a new version of an annotation store\cr
\link[paws.storage:omics_create_multipart_read_set_upload]{create_multipart_read_set_upload} \tab Begins a multipart read set upload\cr
\link[paws.storage:omics_create_reference_store]{create_reference_store} \tab Creates a reference store\cr
\link[paws.storage:omics_create_run_group]{create_run_group} \tab Creates a run group\cr
\link[paws.storage:omics_create_sequence_store]{create_sequence_store} \tab Creates a sequence store\cr
\link[paws.storage:omics_create_share]{create_share} \tab Creates a share offer that can be accepted outside the account by a subscriber\cr
\link[paws.storage:omics_create_variant_store]{create_variant_store} \tab Creates a variant store\cr
\link[paws.storage:omics_create_workflow]{create_workflow} \tab Creates a workflow\cr
\link[paws.storage:omics_delete_annotation_store]{delete_annotation_store} \tab Deletes an annotation store\cr
\link[paws.storage:omics_delete_annotation_store_versions]{delete_annotation_store_versions} \tab Deletes one or multiple versions of an annotation store\cr
\link[paws.storage:omics_delete_reference]{delete_reference} \tab Deletes a genome reference\cr
\link[paws.storage:omics_delete_reference_store]{delete_reference_store} \tab Deletes a genome reference store\cr
\link[paws.storage:omics_delete_run]{delete_run} \tab Deletes a workflow run\cr
\link[paws.storage:omics_delete_run_group]{delete_run_group} \tab Deletes a workflow run group\cr
\link[paws.storage:omics_delete_sequence_store]{delete_sequence_store} \tab Deletes a sequence store\cr
\link[paws.storage:omics_delete_share]{delete_share} \tab Deletes a share of an analytics store\cr
\link[paws.storage:omics_delete_variant_store]{delete_variant_store} \tab Deletes a variant store\cr
\link[paws.storage:omics_delete_workflow]{delete_workflow} \tab Deletes a workflow\cr
\link[paws.storage:omics_get_annotation_import_job]{get_annotation_import_job} \tab Gets information about an annotation import job\cr
\link[paws.storage:omics_get_annotation_store]{get_annotation_store} \tab Gets information about an annotation store\cr
\link[paws.storage:omics_get_annotation_store_version]{get_annotation_store_version} \tab Retrieves the metadata for an annotation store version\cr
\link[paws.storage:omics_get_read_set]{get_read_set} \tab Gets a file from a read set\cr
\link[paws.storage:omics_get_read_set_activation_job]{get_read_set_activation_job} \tab Gets information about a read set activation job\cr
\link[paws.storage:omics_get_read_set_export_job]{get_read_set_export_job} \tab Gets information about a read set export job\cr
\link[paws.storage:omics_get_read_set_import_job]{get_read_set_import_job} \tab Gets information about a read set import job\cr
\link[paws.storage:omics_get_read_set_metadata]{get_read_set_metadata} \tab Gets details about a read set\cr
\link[paws.storage:omics_get_reference]{get_reference} \tab Gets a reference file\cr
\link[paws.storage:omics_get_reference_import_job]{get_reference_import_job} \tab Gets information about a reference import job\cr
\link[paws.storage:omics_get_reference_metadata]{get_reference_metadata} \tab Gets information about a genome reference's metadata\cr
\link[paws.storage:omics_get_reference_store]{get_reference_store} \tab Gets information about a reference store\cr
\link[paws.storage:omics_get_run]{get_run} \tab Gets information about a workflow run\cr
\link[paws.storage:omics_get_run_group]{get_run_group} \tab Gets information about a workflow run group\cr
\link[paws.storage:omics_get_run_task]{get_run_task} \tab Gets information about a workflow run task\cr
\link[paws.storage:omics_get_sequence_store]{get_sequence_store} \tab Gets information about a sequence store\cr
\link[paws.storage:omics_get_share]{get_share} \tab Retrieves the metadata for a share\cr
\link[paws.storage:omics_get_variant_import_job]{get_variant_import_job} \tab Gets information about a variant import job\cr
\link[paws.storage:omics_get_variant_store]{get_variant_store} \tab Gets information about a variant store\cr
\link[paws.storage:omics_get_workflow]{get_workflow} \tab Gets information about a workflow\cr
\link[paws.storage:omics_list_annotation_import_jobs]{list_annotation_import_jobs} \tab Retrieves a list of annotation import jobs\cr
\link[paws.storage:omics_list_annotation_stores]{list_annotation_stores} \tab Retrieves a list of annotation stores\cr
\link[paws.storage:omics_list_annotation_store_versions]{list_annotation_store_versions} \tab Lists the versions of an annotation store\cr
\link[paws.storage:omics_list_multipart_read_set_uploads]{list_multipart_read_set_uploads} \tab Lists multipart read set uploads and for in progress uploads\cr
\link[paws.storage:omics_list_read_set_activation_jobs]{list_read_set_activation_jobs} \tab Retrieves a list of read set activation jobs\cr
\link[paws.storage:omics_list_read_set_export_jobs]{list_read_set_export_jobs} \tab Retrieves a list of read set export jobs\cr
\link[paws.storage:omics_list_read_set_import_jobs]{list_read_set_import_jobs} \tab Retrieves a list of read set import jobs\cr
\link[paws.storage:omics_list_read_sets]{list_read_sets} \tab Retrieves a list of read sets\cr
\link[paws.storage:omics_list_read_set_upload_parts]{list_read_set_upload_parts} \tab This operation will list all parts in a requested multipart upload for a sequence store\cr
\link[paws.storage:omics_list_reference_import_jobs]{list_reference_import_jobs} \tab Retrieves a list of reference import jobs\cr
\link[paws.storage:omics_list_references]{list_references} \tab Retrieves a list of references\cr
\link[paws.storage:omics_list_reference_stores]{list_reference_stores} \tab Retrieves a list of reference stores\cr
\link[paws.storage:omics_list_run_groups]{list_run_groups} \tab Retrieves a list of run groups\cr
\link[paws.storage:omics_list_runs]{list_runs} \tab Retrieves a list of runs\cr
\link[paws.storage:omics_list_run_tasks]{list_run_tasks} \tab Retrieves a list of tasks for a run\cr
\link[paws.storage:omics_list_sequence_stores]{list_sequence_stores} \tab Retrieves a list of sequence stores\cr
\link[paws.storage:omics_list_shares]{list_shares} \tab Lists all shares associated with an account\cr
\link[paws.storage:omics_list_tags_for_resource]{list_tags_for_resource} \tab Retrieves a list of tags for a resource\cr
\link[paws.storage:omics_list_variant_import_jobs]{list_variant_import_jobs} \tab Retrieves a list of variant import jobs\cr
\link[paws.storage:omics_list_variant_stores]{list_variant_stores} \tab Retrieves a list of variant stores\cr
\link[paws.storage:omics_list_workflows]{list_workflows} \tab Retrieves a list of workflows\cr
\link[paws.storage:omics_start_annotation_import_job]{start_annotation_import_job} \tab Starts an annotation import job\cr
\link[paws.storage:omics_start_read_set_activation_job]{start_read_set_activation_job} \tab Activates an archived read set\cr
\link[paws.storage:omics_start_read_set_export_job]{start_read_set_export_job} \tab Exports a read set to Amazon S3\cr
\link[paws.storage:omics_start_read_set_import_job]{start_read_set_import_job} \tab Starts a read set import job\cr
\link[paws.storage:omics_start_reference_import_job]{start_reference_import_job} \tab Starts a reference import job\cr
\link[paws.storage:omics_start_run]{start_run} \tab Starts a workflow run\cr
\link[paws.storage:omics_start_variant_import_job]{start_variant_import_job} \tab Starts a variant import job\cr
\link[paws.storage:omics_tag_resource]{tag_resource} \tab Tags a resource\cr
\link[paws.storage:omics_untag_resource]{untag_resource} \tab Removes tags from a resource\cr
\link[paws.storage:omics_update_annotation_store]{update_annotation_store} \tab Updates an annotation store\cr
\link[paws.storage:omics_update_annotation_store_version]{update_annotation_store_version} \tab Updates the description of an annotation store version\cr
\link[paws.storage:omics_update_run_group]{update_run_group} \tab Updates a run group\cr
\link[paws.storage:omics_update_variant_store]{update_variant_store} \tab Updates a variant store\cr
\link[paws.storage:omics_update_workflow]{update_workflow} \tab Updates a workflow\cr
\link[paws.storage:omics_upload_read_set_part]{upload_read_set_part} \tab This operation uploads a specific part of a read set
}
}

\examples{
\dontrun{
svc <- omics()
svc$abort_multipart_read_set_upload(
  Foo = 123
)
}

}
