% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outposts_operations.R
\name{outposts_create_outpost}
\alias{outposts_create_outpost}
\title{Creates an Outpost}
\usage{
outposts_create_outpost(Name, Description, SiteId, AvailabilityZone,
  AvailabilityZoneId)
}
\arguments{
\item{Name}{}

\item{Description}{}

\item{SiteId}{[required]}

\item{AvailabilityZone}{}

\item{AvailabilityZoneId}{}
}
\description{
Creates an Outpost.
}
\section{Request syntax}{
\preformatted{svc$create_outpost(
  Name = "string",
  Description = "string",
  SiteId = "string",
  AvailabilityZone = "string",
  AvailabilityZoneId = "string"
)
}
}

\keyword{internal}
