% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{pain_class}
\alias{pain_class}
\title{Identify pain class based on pain score}
\usage{
pain_class(score, type = c("binary", "trinary"))
}
\arguments{
\item{score}{pain scores generated by \code{\link{pain_score}}}

\item{type}{binary or trinary (ternary) pain classifications}
}
\value{
one or more pain classes
}
\description{
Convenience function to convert pain scores to binary or trinary pain
classifications, e.g. pain / non-pain or pain / low-pain / high-pain. Note
that trinary classifications are only valid for pain models fit with trinary
classes (e.g. Jones et al. 2020). Otherwise the scale of the score is
arbitrary, and the boundary between low and high pain is not scaled to be 1.
}
