% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws_package.R
\docType{package}
\name{paws-package}
\alias{paws}
\alias{paws-package}
\title{paws: Amazon Web Services Software Development Kit}
\description{
Interface to Amazon Web Services <https://aws.amazon.com>,
    including storage, database, and compute services, such as 'Simple Storage 
    Service' ('S3'), 'DynamoDB' 'NoSQL' database, and 'Lambda' 
    functions-as-a-service.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paws-r/paws}
}

}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
