% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_dataset_group}
\alias{personalize_describe_dataset_group}
\title{Describes the given dataset group}
\usage{
personalize_describe_dataset_group(datasetGroupArn)
}
\arguments{
\item{datasetGroupArn}{[required] The Amazon Resource Name (ARN) of the dataset group to describe.}
}
\description{
Describes the given dataset group. For more information on dataset
groups, see CreateDatasetGroup.
}
\section{Request syntax}{
\preformatted{svc$describe_dataset_group(
  datasetGroupArn = "string"
)
}
}

\keyword{internal}
