% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalizeevents_service.R
\name{personalizeevents}
\alias{personalizeevents}
\title{Amazon Personalize Events}
\usage{
personalizeevents()
}
\description{
Amazon Personalize Events
}
\section{Operations}{

\tabular{ll}{
\link[=personalizeevents_put_events]{put_events} \tab Records user interaction event data
}
}

\examples{
\donttest{svc <- personalizeevents()
svc$put_events(
  Foo = 123
)}

}
