% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{personalizeevents}
\alias{personalizeevents}
\title{Amazon Personalize Events}
\usage{
personalizeevents()
}
\arguments{
\item{config}{An optional list of custom configurations for the service. Currently
supports adding custom credentials, endpoint, and region.}
}
\description{
Amazon Personalize Events
}
\section{Service syntax}{
\preformatted{svc <- personalizeevents(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string",
        provider_name = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=personalizeevents_put_events]{put_events} \tab Records user interaction event data
}
}

\examples{
\donttest{svc <- personalizeevents()
svc$put_events(
  Foo = 123
)}

}
