% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{personalizeruntime}
\alias{personalizeruntime}
\title{Amazon Personalize Runtime}
\usage{
personalizeruntime()
}
\description{
Amazon Personalize Runtime
}
\section{Operations}{

\tabular{ll}{
\link[=personalizeruntime_get_personalized_ranking]{get_personalized_ranking} \tab Re-ranks a list of recommended items for the given user\cr
\link[=personalizeruntime_get_recommendations]{get_recommendations} \tab Returns a list of recommended items
}
}

\examples{
\donttest{svc <- personalizeruntime()
svc$get_personalized_ranking(
  Foo = 123
)}

}
