% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_segment}
\alias{pinpoint_delete_segment}
\title{Deletes a segment from an application}
\usage{
pinpoint_delete_segment(ApplicationId, SegmentId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{SegmentId}{[required] The unique identifier for the segment.}
}
\description{
Deletes a segment from an application.
}
\section{Request syntax}{
\preformatted{svc$delete_segment(
  ApplicationId = "string",
  SegmentId = "string"
)
}
}

\keyword{internal}
