% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_application_settings}
\alias{pinpoint_get_application_settings}
\title{Used to request the settings for an app}
\usage{
pinpoint_get_application_settings(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Used to request the settings for an app.
}
\section{Request syntax}{
\preformatted{pinpoint$get_application_settings(
  ApplicationId = "string"
)
}
}

\keyword{internal}
