% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_channels}
\alias{pinpoint_get_channels}
\title{Get all channels}
\usage{
pinpoint_get_channels(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get all channels.
}
\section{Request syntax}{
\preformatted{svc$get_channels(
  ApplicationId = "string"
)
}
}

\keyword{internal}
