% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_gcm_channel}
\alias{pinpoint_update_gcm_channel}
\title{Use to update the GCM channel for an app}
\usage{
pinpoint_update_gcm_channel(ApplicationId, GCMChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{GCMChannelRequest}{[required]}
}
\description{
Use to update the GCM channel for an app.
}
\section{Request syntax}{
\preformatted{pinpoint$update_gcm_channel(
  ApplicationId = "string",
  GCMChannelRequest = list(
    ApiKey = "string",
    Enabled = TRUE|FALSE
  )
)
}
}

\keyword{internal}
