% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_service.R
\name{prometheusservice}
\alias{prometheusservice}
\title{Amazon Prometheus Service}
\usage{
prometheusservice(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Managed Service for Prometheus
}
\section{Service syntax}{
\preformatted{svc <- prometheusservice(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=prometheusservice_create_workspace]{create_workspace} \tab Creates a new AMP workspace\cr
\link[=prometheusservice_delete_workspace]{delete_workspace} \tab Deletes an AMP workspace\cr
\link[=prometheusservice_describe_workspace]{describe_workspace} \tab Describes an existing AMP workspace\cr
\link[=prometheusservice_list_workspaces]{list_workspaces} \tab Lists all AMP workspaces, including workspaces being created or deleted\cr
\link[=prometheusservice_update_workspace_alias]{update_workspace_alias} \tab Updates an AMP workspace alias
}
}

\examples{
\dontrun{
svc <- prometheusservice()
svc$create_workspace(
  Foo = 123
)
}

}
