% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldbsession_service.R
\name{qldbsession}
\alias{qldbsession}
\title{Amazon QLDB Session}
\usage{
qldbsession(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
The transactional data APIs for Amazon QLDB

Instead of interacting directly with this API, we recommend that you use
the Amazon QLDB Driver or the QLDB Shell to execute data transactions on
a ledger.
\itemize{
\item If you are working with an AWS SDK, use the QLDB Driver. The driver
provides a high-level abstraction layer above this \code{qldbsession}
data plane and manages \code{SendCommand} API calls for you. For
information and a list of supported programming languages, see
\href{https://docs.aws.amazon.com/qldb/latest/developerguide/getting-started-driver.html}{Getting started with the driver}
in the \emph{Amazon QLDB Developer Guide}.
\item If you are working with the AWS Command Line Interface (AWS CLI),
use the QLDB Shell. The shell is a command line interface that uses
the QLDB Driver to interact with a ledger. For information, see
\href{https://docs.aws.amazon.com/qldb/latest/developerguide/data-shell.html}{Accessing Amazon QLDB using the QLDB Shell}.
}
}
\section{Service syntax}{
\preformatted{svc <- qldbsession(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=qldbsession_send_command]{send_command} \tab Sends a command to an Amazon QLDB ledger
}
}

\examples{
\dontrun{
svc <- qldbsession()
svc$send_command(
  Foo = 123
)
}

}
