% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_dashboard}
\alias{quicksight_delete_dashboard}
\title{Deletes a dashboard}
\usage{
quicksight_delete_dashboard(AwsAccountId, DashboardId, VersionNumber)
}
\arguments{
\item{AwsAccountId}{[required] AWS account ID that contains the dashboard you are deleting.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{VersionNumber}{The version number of the dashboard. If version number property is
provided, only the specified version of the dashboard is deleted.}
}
\description{
Deletes a dashboard.
}
\section{Request syntax}{
\preformatted{svc$delete_dashboard(
  AwsAccountId = "string",
  DashboardId = "string",
  VersionNumber = 123
)
}
}

\keyword{internal}
