% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_template_alias}
\alias{quicksight_describe_template_alias}
\title{Describes the template aliases of a template}
\usage{
quicksight_describe_template_alias(AwsAccountId, TemplateId, AliasName)
}
\arguments{
\item{AwsAccountId}{[required] AWS account ID that contains the template alias you are describing.}

\item{TemplateId}{[required] An ID for the template.}

\item{AliasName}{[required] The alias of the template that you want to describe. If you provide a
specific alias, you describe the version that the alias points to. You
can specify the latest version of the template by providing the keyword
\verb{$LATEST} in the \code{AliasName} parameter. The keyword \verb{$PUBLISHED}
doesn\'t apply to templates.}
}
\description{
Describes the template aliases of a template.
}
\section{Request syntax}{
\preformatted{svc$describe_template_alias(
  AwsAccountId = "string",
  TemplateId = "string",
  AliasName = "string"
)
}
}

\keyword{internal}
