% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_register_db_proxy_targets}
\alias{rds_register_db_proxy_targets}
\title{This is prerelease documentation for the RDS Database Proxy feature in
preview release}
\usage{
rds_register_db_proxy_targets(DBProxyName, TargetGroupName,
  DBInstanceIdentifiers, DBClusterIdentifiers)
}
\arguments{
\item{DBProxyName}{[required] The identifier of the \code{DBProxy} that is associated with the
\code{DBProxyTargetGroup}.}

\item{TargetGroupName}{The identifier of the \code{DBProxyTargetGroup}.}

\item{DBInstanceIdentifiers}{One or more DB instance identifiers.}

\item{DBClusterIdentifiers}{One or more DB cluster identifiers.}
}
\description{
This is prerelease documentation for the RDS Database Proxy feature in
preview release. It is subject to change.
}
\details{
Associate one or more \code{DBProxyTarget} data structures with a
\code{DBProxyTargetGroup}.
}
\section{Request syntax}{
\preformatted{svc$register_db_proxy_targets(
  DBProxyName = "string",
  TargetGroupName = "string",
  DBInstanceIdentifiers = list(
    "string"
  ),
  DBClusterIdentifiers = list(
    "string"
  )
)
}
}

\keyword{internal}
