% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_copy_object}
\alias{s3_copy_object}
\title{Creates a copy of an object that is already stored in Amazon S3}
\usage{
s3_copy_object(ACL, Bucket, CacheControl, ContentDisposition,
  ContentEncoding, ContentLanguage, ContentType, CopySource,
  CopySourceIfMatch, CopySourceIfModifiedSince, CopySourceIfNoneMatch,
  CopySourceIfUnmodifiedSince, Expires, GrantFullControl, GrantRead,
  GrantReadACP, GrantWriteACP, Key, Metadata, MetadataDirective,
  TaggingDirective, ServerSideEncryption, StorageClass,
  WebsiteRedirectLocation, SSECustomerAlgorithm, SSECustomerKey,
  SSECustomerKeyMD5, SSEKMSKeyId, SSEKMSEncryptionContext,
  CopySourceSSECustomerAlgorithm, CopySourceSSECustomerKey,
  CopySourceSSECustomerKeyMD5, RequestPayer, Tagging, ObjectLockMode,
  ObjectLockRetainUntilDate, ObjectLockLegalHoldStatus)
}
\arguments{
\item{ACL}{The canned ACL to apply to the object.}

\item{Bucket}{[required] The name of the destination bucket.}

\item{CacheControl}{Specifies caching behavior along the request/reply chain.}

\item{ContentDisposition}{Specifies presentational information for the object.}

\item{ContentEncoding}{Specifies what content encodings have been applied to the object and
thus what decoding mechanisms must be applied to obtain the media-type
referenced by the Content-Type header field.}

\item{ContentLanguage}{The language the content is in.}

\item{ContentType}{A standard MIME type describing the format of the object data.}

\item{CopySource}{[required] The name of the source bucket and key name of the source object,
separated by a slash (/). Must be URL-encoded.}

\item{CopySourceIfMatch}{Copies the object if its entity tag (ETag) matches the specified tag.}

\item{CopySourceIfModifiedSince}{Copies the object if it has been modified since the specified time.}

\item{CopySourceIfNoneMatch}{Copies the object if its entity tag (ETag) is different than the
specified ETag.}

\item{CopySourceIfUnmodifiedSince}{Copies the object if it hasn\'t been modified since the specified time.}

\item{Expires}{The date and time at which the object is no longer cacheable.}

\item{GrantFullControl}{Gives the grantee READ, READ\\_ACP, and WRITE\\_ACP permissions on the
object.}

\item{GrantRead}{Allows grantee to read the object data and its metadata.}

\item{GrantReadACP}{Allows grantee to read the object ACL.}

\item{GrantWriteACP}{Allows grantee to write the ACL for the applicable object.}

\item{Key}{[required] The key of the destination object.}

\item{Metadata}{A map of metadata to store with the object in S3.}

\item{MetadataDirective}{Specifies whether the metadata is copied from the source object or
replaced with metadata provided in the request.}

\item{TaggingDirective}{Specifies whether the object tag-set are copied from the source object
or replaced with tag-set provided in the request.}

\item{ServerSideEncryption}{The server-side encryption algorithm used when storing this object in
Amazon S3 (for example, AES256, aws:kms).}

\item{StorageClass}{The type of storage to use for the object. Defaults to \'STANDARD\'.}

\item{WebsiteRedirectLocation}{If the bucket is configured as a website, redirects requests for this
object to another object in the same bucket or to an external URL.
Amazon S3 stores the value of this header in the object metadata.}

\item{SSECustomerAlgorithm}{Specifies the algorithm to use to when encrypting the object (for
example, AES256).}

\item{SSECustomerKey}{Specifies the customer-provided encryption key for Amazon S3 to use in
encrypting data. This value is used to store the object and then it is
discarded; Amazon S3 does not store the encryption key. The key must be
appropriate for use with the algorithm specified in the
\verb{x-amz-server-side​-encryption​-customer-algorithm} header.}

\item{SSECustomerKeyMD5}{Specifies the 128-bit MD5 digest of the encryption key according to RFC
1321. Amazon S3 uses this header for a message integrity check to ensure
that the encryption key was transmitted without error.}

\item{SSEKMSKeyId}{Specifies the AWS KMS key ID to use for object encryption. All GET and
PUT requests for an object protected by AWS KMS will fail if not made
via SSL or using SigV4. For information about configuring using any of
the officially supported AWS SDKs and AWS CLI, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version}{Specifying the Signature Version in Request Authentication}
in the \emph{Amazon S3 Developer Guide}.}

\item{SSEKMSEncryptionContext}{Specifies the AWS KMS Encryption Context to use for object encryption.
The value of this header is a base64-encoded UTF-8 string holding JSON
with the encryption context key-value pairs.}

\item{CopySourceSSECustomerAlgorithm}{Specifies the algorithm to use when decrypting the source object (for
example, AES256).}

\item{CopySourceSSECustomerKey}{Specifies the customer-provided encryption key for Amazon S3 to use to
decrypt the source object. The encryption key provided in this header
must be one that was used when the source object was created.}

\item{CopySourceSSECustomerKeyMD5}{Specifies the 128-bit MD5 digest of the encryption key according to RFC
1321. Amazon S3 uses this header for a message integrity check to ensure
that the encryption key was transmitted without error.}

\item{RequestPayer}{}

\item{Tagging}{The tag-set for the object destination object this value must be used in
conjunction with the \code{TaggingDirective}. The tag-set must be encoded as
URL Query parameters.}

\item{ObjectLockMode}{The Object Lock mode that you want to apply to the copied object.}

\item{ObjectLockRetainUntilDate}{The date and time when you want the copied object\'s Object Lock to
expire.}

\item{ObjectLockLegalHoldStatus}{Specifies whether you want to apply a Legal Hold to the copied object.}
}
\description{
Creates a copy of an object that is already stored in Amazon S3.
}
\details{
You can store individual objects of up to 5 TB in Amazon S3. You create
a copy of your object up to 5 GB in size in a single atomic operation
using this API. However, for copying an object greater than 5 GB, you
must use the multipart upload Upload Part - Copy API. For more
information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/CopyingObjctsUsingRESTMPUapi.html}{Copy Object Using the REST Multipart Upload API}.

When copying an object, you can preserve all metadata (default) or
specify new metadata. However, the ACL is not preserved and is set to
private for the user making the request. To override the default ACL
setting, specify a new ACL when generating a copy request. For more
information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/S3_ACLs_UsingACLs.html}{Using ACLs}.

Amazon S3 transfer acceleration does not support cross-region copies. If
you request a cross-region copy using a transfer acceleration endpoint,
you get a 400 \verb{Bad Request} error. For more information about transfer
acceleration, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/transfer-acceleration.html}{Transfer Acceleration}.

All copy requests must be authenticated. Additionally, you must have
\emph{read} access to the source object and \emph{write} access to the destination
bucket. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/RESTAuthentication.html}{REST Authentication}.
Both the Region that you want to copy the object from and the Region
that you want to copy the object to must be enabled for your account.

To only copy an object under certain conditions, such as whether the
\code{Etag} matches or whether the object was modified before or after a
specified date, use the request parameters \verb{x-amz-copy-source-if-match},
\verb{x-amz-copy-source-if-none-match},
\verb{x-amz-copy-source-if-unmodified-since}, or
\verb{ x-amz-copy-source-if-modified-since}.

All headers with the \verb{x-amz-} prefix, including \code{x-amz-copy-source},
must be signed.

You can use this operation to change the storage class of an object that
is already stored in Amazon S3 using the \code{StorageClass} parameter. For
more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/storage-class-intro.html}{Storage Classes}.

The source object that you are copying can be encrypted or unencrypted.
If the source object is encrypted, it can be encrypted by server-side
encryption using AWS managed encryption keys or by using a
customer-provided encryption key. When copying an object, you can
request that Amazon S3 encrypt the target object by using either the AWS
managed encryption keys or by using your own encryption key. You can do
this regardless of the form of server-side encryption that was used to
encrypt the source, or even if the source object was not encrypted. For
more information about server-side encryption, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/serv-side-encryption.html}{Using Server-Side Encryption}.

A copy request might return an error when Amazon S3 receives the copy
request or while Amazon S3 is copying the files. If the error occurs
before the copy operation starts, you receive a standard Amazon S3
error. If the error occurs during the copy operation, the error response
is embedded in the \verb{200 OK} response. This means that a \verb{200 OK}
response can contain either a success or an error. Design your
application to parse the contents of the response and handle it
appropriately.

If the copy is successful, you receive a response with information about
the copied object.

If the request is an HTTP 1.1 request, the response is chunk encoded. If
it were not, it would not contain the content-length, and you would need
to read the entire body.

Consider the following when using request headers:
\itemize{
\item Consideration 1 -- If both the \verb{x-amz-copy-source-if-match} and
\verb{x-amz-copy-source-if-unmodified-since} headers are present in the
request and evaluate as follows, Amazon S3 returns 200 OK and copies
the data:
\itemize{
\item \verb{x-amz-copy-source-if-match} condition evaluates to true
\item \verb{x-amz-copy-source-if-unmodified-since} condition evaluates to
false
}
\item Consideration 2 -- If both of the \verb{x-amz-copy-source-if-none-match}
and \verb{x-amz-copy-source-if-modified-since} headers are present in the
request and evaluate as follows, Amazon S3 returns the
\verb{412 Precondition Failed} response code:
\itemize{
\item \verb{x-amz-copy-source-if-none-match} condition evaluates to false
\item \verb{x-amz-copy-source-if-modified-since} condition evaluates to
true
}
}

The copy request charge is based on the storage class and Region you
specify for the destination object. For pricing information, see \href{https://aws.amazon.com/s3/pricing/}{Amazon S3 Pricing}.

Following are other considerations when using \code{CopyObject}:
\subsection{Versioning}{

By default, \code{x-amz-copy-source} identifies the current version of an
object to copy. (If the current version is a delete marker, Amazon S3
behaves as if the object was deleted.) To copy a different version, use
the \code{versionId} subresource.

If you enable versioning on the target bucket, Amazon S3 generates a
unique version ID for the object being copied. This version ID is
different from the version ID of the source object. Amazon S3 returns
the version ID of the copied object in the \code{x-amz-version-id} response
header in the response.

If you do not enable versioning or suspend it on the target bucket, the
version ID that Amazon S3 generates is always null.

If the source object\'s storage class is GLACIER, you must restore a
copy of this object before you can use it as a source object for the
copy operation. For more information, see .
}

\subsection{Access Permissions}{

When copying an object, you can optionally specify the accounts or
groups that should be granted specific permissions on the new object.
There are two ways to grant the permissions using the request headers:
\itemize{
\item Specify a canned ACL with the \code{x-amz-acl} request header. For more
information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#CannedACL}{Canned ACL}.
\item Specify access permissions explicitly with the \code{x-amz-grant-read},
\code{x-amz-grant-read-acp}, \code{x-amz-grant-write-acp}, and
\code{x-amz-grant-full-control} headers. These parameters map to the set
of permissions that Amazon S3 supports in an ACL. For more
information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html}{Access Control List (ACL) Overview}.
}

You can use either a canned ACL or specify access permissions
explicitly. You cannot do both.
}

\subsection{Server-Side- Encryption-Specific Request Headers}{

To encrypt the target object, you must provide the appropriate
encryption-related request headers. The one you use depends on whether
you want to use AWS managed encryption keys or provide your own
encryption key.
\itemize{
\item To encrypt the target object using server-side encryption with an
AWS managed encryption key, provide the following request headers,
as appropriate.
\itemize{
\item \verb{x-amz-server-side​-encryption}
\item \code{x-amz-server-side-encryption-aws-kms-key-id}
\item \code{x-amz-server-side-encryption-context}
}

If you specify \code{x-amz-server-side-encryption:aws:kms}, but don\'t
provide \code{x-amz-server-side-encryption-aws-kms-key-id}, Amazon S3
uses the AWS managed CMK in AWS KMS to protect the data. If you want
to use a customer managed AWS KMS CMK, you must provide the
\code{x-amz-server-side-encryption-aws-kms-key-id} of the symmetric
customer managed CMK. Amazon S3 only supports symmetric CMKs and not
asymmetric CMKs. For more information, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/symmetric-asymmetric.html}{Using Symmetric and Asymmetric Keys}
in the \emph{AWS Key Management Service Developer Guide}.

All GET and PUT requests for an object protected by AWS KMS fail if
you don\'t make them with SSL or by using SigV4.

For more information about server-side encryption with CMKs stored
in AWS KMS (SSE-KMS), see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/UsingKMSEncryption.html}{Protecting Data Using Server-Side Encryption with CMKs stored in KMS}.
\item To encrypt the target object using server-side encryption with an
encryption key that you provide, use the following headers.
\itemize{
\item x-amz-server-side​-encryption​-customer-algorithm
\item x-amz-server-side​-encryption​-customer-key
\item x-amz-server-side​-encryption​-customer-key-MD5
}
\item If the source object is encrypted using server-side encryption with
customer-provided encryption keys, you must use the following
headers.
\itemize{
\item x-amz-copy-source​-server-side​-encryption​-customer-algorithm
\item x-amz-copy-source​-server-side​-encryption​-customer-key
\item x-amz-copy-source-​server-side​-encryption​-customer-key-MD5
}

For more information about server-side encryption with CMKs stored
in AWS KMS (SSE-KMS), see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/UsingKMSEncryption.html}{Protecting Data Using Server-Side Encryption with CMKs stored in Amazon KMS}.
}
}

\subsection{Access-Control-List (ACL)-Specific Request Headers}{

You also can use the following access control--related headers with this
operation. By default, all objects are private. Only the owner has full
access control. When adding a new object, you can grant permissions to
individual AWS accounts or to predefined groups defined by Amazon S3.
These permissions are then added to the access control list (ACL) on the
object. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/S3_ACLs_UsingACLs.html}{Using ACLs}.
With this operation, you can grant access permissions using one of the
following two methods:
\itemize{
\item Specify a canned ACL (\code{x-amz-acl}) --- Amazon S3 supports a set of
predefined ACLs, known as \emph{canned ACLs}. Each canned ACL has a
predefined set of grantees and permissions. For more information,
see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#CannedACL}{Canned ACL}.
\item Specify access permissions explicitly --- To explicitly grant access
permissions to specific AWS accounts or groups, use the following
headers. Each header maps to specific permissions that Amazon S3
supports in an ACL. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html}{Access Control List (ACL) Overview}.
In the header, you specify a list of grantees who get the specific
permission. To grant permissions explicitly, use:
\itemize{
\item x-amz-grant-read
\item x-amz-grant-write
\item x-amz-grant-read-acp
\item x-amz-grant-write-acp
\item x-amz-grant-full-control
}

You specify each grantee as a type=value pair, where the type is one
of the following:
\itemize{
\item \code{emailAddress} -- if the value specified is the email address of
an AWS account
\item \code{id} -- if the value specified is the canonical user ID of an
AWS account
\item \code{uri} -- if you are granting permissions to a predefined group
}

For example, the following \code{x-amz-grant-read} header grants the AWS
accounts identified by email addresses permissions to read object
data and its metadata:

\verb{x-amz-grant-read: emailAddress="xyz@amazon.com", emailAddress="abc@amazon.com" }
}

The following operations are related to \code{CopyObject}:
\itemize{
\item PutObject
\item GetObject
}

For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/CopyingObjectsExamples.html}{Copying Objects}.
}
}
\section{Request syntax}{
\preformatted{svc$copy_object(
  ACL = "private"|"public-read"|"public-read-write"|"authenticated-read"|"aws-exec-read"|"bucket-owner-read"|"bucket-owner-full-control",
  Bucket = "string",
  CacheControl = "string",
  ContentDisposition = "string",
  ContentEncoding = "string",
  ContentLanguage = "string",
  ContentType = "string",
  CopySource = "string",
  CopySourceIfMatch = "string",
  CopySourceIfModifiedSince = as.POSIXct(
    "2015-01-01"
  ),
  CopySourceIfNoneMatch = "string",
  CopySourceIfUnmodifiedSince = as.POSIXct(
    "2015-01-01"
  ),
  Expires = as.POSIXct(
    "2015-01-01"
  ),
  GrantFullControl = "string",
  GrantRead = "string",
  GrantReadACP = "string",
  GrantWriteACP = "string",
  Key = "string",
  Metadata = list(
    "string"
  ),
  MetadataDirective = "COPY"|"REPLACE",
  TaggingDirective = "COPY"|"REPLACE",
  ServerSideEncryption = "AES256"|"aws:kms",
  StorageClass = "STANDARD"|"REDUCED_REDUNDANCY"|"STANDARD_IA"|"ONEZONE_IA"|"INTELLIGENT_TIERING"|"GLACIER"|"DEEP_ARCHIVE",
  WebsiteRedirectLocation = "string",
  SSECustomerAlgorithm = "string",
  SSECustomerKey = raw,
  SSECustomerKeyMD5 = "string",
  SSEKMSKeyId = "string",
  SSEKMSEncryptionContext = "string",
  CopySourceSSECustomerAlgorithm = "string",
  CopySourceSSECustomerKey = raw,
  CopySourceSSECustomerKeyMD5 = "string",
  RequestPayer = "requester",
  Tagging = "string",
  ObjectLockMode = "GOVERNANCE"|"COMPLIANCE",
  ObjectLockRetainUntilDate = as.POSIXct(
    "2015-01-01"
  ),
  ObjectLockLegalHoldStatus = "ON"|"OFF"
)
}
}

\examples{
# The following example copies an object from one bucket to another.
\dontrun{svc$copy_object(
  Bucket = "destinationbucket",
  CopySource = "/sourcebucket/HappyFacejpg",
  Key = "HappyFaceCopyjpg"
)}

}
\keyword{internal}
