% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_delete_public_access_block}
\alias{s3control_delete_public_access_block}
\title{Deletes the block public access configuration for the specified account}
\usage{
s3control_delete_public_access_block(AccountId)
}
\arguments{
\item{AccountId}{[required] The account ID for the AWS account whose block public access
configuration you want to delete.}
}
\description{
Deletes the block public access configuration for the specified account.
}
\section{Request syntax}{
\preformatted{svc$delete_public_access_block(
  AccountId = "string"
)
}
}

\keyword{internal}
