% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3outposts_service.R
\name{s3outposts}
\alias{s3outposts}
\title{Amazon S3 on Outposts}
\usage{
s3outposts(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon S3 on Outposts provides access to S3 on Outposts operations.
}
\section{Service syntax}{
\preformatted{svc <- s3outposts(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=s3outposts_create_endpoint]{create_endpoint} \tab S3 on Outposts access points simplify managing data access at scale for shared datasets in Amazon S3 on Outposts\cr
\link[=s3outposts_delete_endpoint]{delete_endpoint} \tab S3 on Outposts access points simplify managing data access at scale for shared datasets in Amazon S3 on Outposts\cr
\link[=s3outposts_list_endpoints]{list_endpoints} \tab S3 on Outposts access points simplify managing data access at scale for shared datasets in Amazon S3 on Outposts
}
}

\examples{
\dontrun{
svc <- s3outposts()
svc$create_endpoint(
  Foo = 123
)
}

}
