% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3outposts_operations.R
\name{s3outposts_create_endpoint}
\alias{s3outposts_create_endpoint}
\title{S3 on Outposts access points simplify managing data access at scale for
shared datasets in Amazon S3 on Outposts}
\usage{
s3outposts_create_endpoint(OutpostId, SubnetId, SecurityGroupId)
}
\arguments{
\item{OutpostId}{[required] The ID of the AWS Outpost.}

\item{SubnetId}{[required] The ID of the subnet in the selected VPC.}

\item{SecurityGroupId}{[required] The ID of the security group to use with the endpoint.}
}
\description{
S3 on Outposts access points simplify managing data access at scale for
shared datasets in Amazon S3 on Outposts. S3 on Outposts uses endpoints
to connect to Outposts buckets so that you can perform actions within
your virtual private cloud (VPC).

This action creates an endpoint and associates it with the specified
Outpost.

Related actions include:
\itemize{
\item \code{\link[=s3outposts_delete_endpoint]{delete_endpoint}}
\item \code{\link[=s3outposts_list_endpoints]{list_endpoints}}
}
}
\section{Request syntax}{
\preformatted{svc$create_endpoint(
  OutpostId = "string",
  SubnetId = "string",
  SecurityGroupId = "string"
)
}
}

\keyword{internal}
