% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3outposts_operations.R
\name{s3outposts_delete_endpoint}
\alias{s3outposts_delete_endpoint}
\title{S3 on Outposts access points simplify managing data access at scale for
shared datasets in Amazon S3 on Outposts}
\usage{
s3outposts_delete_endpoint(EndpointId, OutpostId)
}
\arguments{
\item{EndpointId}{[required] The ID of the end point.}

\item{OutpostId}{[required] The ID of the AWS Outpost.}
}
\description{
S3 on Outposts access points simplify managing data access at scale for
shared datasets in Amazon S3 on Outposts. S3 on Outposts uses endpoints
to connect to Outposts buckets so that you can perform actions within
your virtual private cloud (VPC).

This action deletes an endpoint.

Related actions include:
\itemize{
\item \code{\link[=s3outposts_create_endpoint]{create_endpoint}}
\item \code{\link[=s3outposts_list_endpoints]{list_endpoints}}
}
}
\section{Request syntax}{
\preformatted{svc$delete_endpoint(
  EndpointId = "string",
  OutpostId = "string"
)
}
}

\keyword{internal}
