% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_user_profile}
\alias{sagemaker_describe_user_profile}
\title{Describes a user profile}
\usage{
sagemaker_describe_user_profile(DomainId, UserProfileName)
}
\arguments{
\item{DomainId}{[required] The domain ID.}

\item{UserProfileName}{[required] The user profile name.}
}
\description{
Describes a user profile. For more information, see
\code{\link[=sagemaker_create_user_profile]{create_user_profile}}.
}
\section{Request syntax}{
\preformatted{svc$describe_user_profile(
  DomainId = "string",
  UserProfileName = "string"
)
}
}

\keyword{internal}
