% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakeredgemanager_service.R
\name{sagemakeredgemanager}
\alias{sagemakeredgemanager}
\title{Amazon Sagemaker Edge Manager}
\usage{
sagemakeredgemanager(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
SageMaker Edge Manager dataplane service for communicating with active
agents.
}
\section{Service syntax}{
\preformatted{svc <- sagemakeredgemanager(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=sagemakeredgemanager_get_device_registration]{get_device_registration} \tab Use to check if a device is registered with SageMaker Edge Manager\cr
\link[=sagemakeredgemanager_send_heartbeat]{send_heartbeat} \tab Use to get the current status of devices registered on SageMaker Edge Manager
}
}

\examples{
\dontrun{
svc <- sagemakeredgemanager()
svc$get_device_registration(
  Foo = 123
)
}

}
