% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_service.R
\name{savingsplans}
\alias{savingsplans}
\title{AWS Savings Plans}
\usage{
savingsplans(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Savings Plans are a pricing model that offer significant savings on AWS
usage (for example, on Amazon EC2 instances). You commit to a consistent
amount of usage, in USD per hour, for a term of 1 or 3 years, and
receive a lower price for that usage. For more information, see the \href{https://docs.aws.amazon.com/savingsplans/latest/userguide/}{AWS Savings Plans User Guide}.
}
\section{Service syntax}{
\preformatted{svc <- savingsplans(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=savingsplans_create_savings_plan]{create_savings_plan} \tab Creates a Savings Plan \cr
\link[=savingsplans_describe_savings_plan_rates]{describe_savings_plan_rates} \tab Describes the specified Savings Plans rates \cr
\link[=savingsplans_describe_savings_plans]{describe_savings_plans} \tab Describes the specified Savings Plans \cr
\link[=savingsplans_describe_savings_plans_offering_rates]{describe_savings_plans_offering_rates} \tab Describes the specified Savings Plans offering rates \cr
\link[=savingsplans_describe_savings_plans_offerings]{describe_savings_plans_offerings} \tab Describes the specified Savings Plans offerings \cr
\link[=savingsplans_list_tags_for_resource]{list_tags_for_resource} \tab Lists the tags for the specified resource \cr
\link[=savingsplans_tag_resource]{tag_resource} \tab Adds the specified tags to the specified resource \cr
\link[=savingsplans_untag_resource]{untag_resource} \tab Removes the specified tags from the specified resource
}
}

\examples{
\dontrun{svc <- savingsplans()
svc$create_savings_plan(
  Foo = 123
)}

}
