% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_delete_schema_version}
\alias{schemas_delete_schema_version}
\title{Delete the schema version definition}
\usage{
schemas_delete_schema_version(RegistryName, SchemaName, SchemaVersion)
}
\arguments{
\item{RegistryName}{[required]}

\item{SchemaName}{[required]}

\item{SchemaVersion}{[required]}
}
\description{
Delete the schema version definition
}
\section{Request syntax}{
\preformatted{svc$delete_schema_version(
  RegistryName = "string",
  SchemaName = "string",
  SchemaVersion = "string"
)
}
}

\keyword{internal}
