% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_list_discoverers}
\alias{schemas_list_discoverers}
\title{List the discoverers}
\usage{
schemas_list_discoverers(DiscovererIdPrefix, Limit, NextToken,
  SourceArnPrefix)
}
\arguments{
\item{DiscovererIdPrefix}{}

\item{Limit}{}

\item{NextToken}{}

\item{SourceArnPrefix}{}
}
\description{
List the discoverers.
}
\section{Request syntax}{
\preformatted{svc$list_discoverers(
  DiscovererIdPrefix = "string",
  Limit = 123,
  NextToken = "string",
  SourceArnPrefix = "string"
)
}
}

\keyword{internal}
