% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_create_members}
\alias{securityhub_create_members}
\title{Creates member Security Hub accounts in the current AWS account (which
becomes the master Security Hub account) that has Security Hub enabled}
\usage{
securityhub_create_members(AccountDetails)
}
\arguments{
\item{AccountDetails}{A list of account ID and email address pairs of the accounts that you
want to associate with the master Security Hub account.}
}
\description{
Creates member Security Hub accounts in the current AWS account (which
becomes the master Security Hub account) that has Security Hub enabled.
}
\section{Request syntax}{
\preformatted{securityhub$create_members(
  AccountDetails = list(
    list(
      AccountId = "string",
      Email = "string"
    )
  )
)
}
}

\keyword{internal}
