% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_decline_invitations}
\alias{securityhub_decline_invitations}
\title{Declines invitations to become a member account}
\usage{
securityhub_decline_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{[required] A list of account IDs that specify the accounts that invitations to
Security Hub are declined from.}
}
\description{
Declines invitations to become a member account.
}
\section{Request syntax}{
\preformatted{svc$decline_invitations(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
