% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disassociate_members}
\alias{securityhub_disassociate_members}
\title{Disassociates the specified member accounts from the associated master
account}
\usage{
securityhub_disassociate_members(AccountIds)
}
\arguments{
\item{AccountIds}{The account IDs of the member accounts to disassociate from the master
account.}
}
\description{
Disassociates the specified member accounts from the associated master
account.
}
\section{Request syntax}{
\preformatted{svc$disassociate_members(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
