% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_invitations}
\alias{securityhub_list_invitations}
\title{Lists all Security Hub membership invitations that were sent to the
current AWS account}
\usage{
securityhub_list_invitations(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of items to return in the response.}

\item{NextToken}{The token that is required for pagination. On your first call to the
\code{\link[=securityhub_list_invitations]{list_invitations}} operation, set the
value of this parameter to \code{NULL}.

For subsequent calls to the operation, to continue listing data, set the
value of this parameter to the value returned from the previous
response.}
}
\description{
Lists all Security Hub membership invitations that were sent to the
current AWS account.

This operation is only used by accounts that do not belong to an
organization. Organization accounts do not receive invitations.
}
\section{Request syntax}{
\preformatted{svc$list_invitations(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
