% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_provisioned_product}
\alias{servicecatalog_describe_provisioned_product}
\title{Gets information about the specified provisioned product}
\usage{
servicecatalog_describe_provisioned_product(AcceptLanguage, Id)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The provisioned product identifier.}
}
\description{
Gets information about the specified provisioned product.
}
\section{Request syntax}{
\preformatted{servicecatalog$describe_provisioned_product(
  AcceptLanguage = "string",
  Id = "string"
)
}
}

\keyword{internal}
