% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{servicediscovery}
\alias{servicediscovery}
\title{AWS Cloud Map}
\usage{
servicediscovery(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
AWS Cloud Map lets you configure public DNS, private DNS, or HTTP
namespaces that your microservice applications run in. When an instance
of the service becomes available, you can call the AWS Cloud Map API to
register the instance with AWS Cloud Map. For public or private DNS
namespaces, AWS Cloud Map automatically creates DNS records and an
optional health check. Clients that submit public or private DNS
queries, or HTTP requests, for the service receive an answer that
contains up to eight healthy records.
}
\section{Service syntax}{
\preformatted{svc <- servicediscovery(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=servicediscovery_create_http_namespace]{create_http_namespace} \tab Creates an HTTP namespace \cr
\link[=servicediscovery_create_private_dns_namespace]{create_private_dns_namespace} \tab Creates a private namespace based on DNS, which will be visible only inside a specified Amazon VPC \cr
\link[=servicediscovery_create_public_dns_namespace]{create_public_dns_namespace} \tab Creates a public namespace based on DNS, which will be visible on the internet \cr
\link[=servicediscovery_create_service]{create_service} \tab Creates a service, which defines the configuration for the following entities: - For public and private DNS namespaces, one of the following combinations of DNS records in Amazon Route 53: - A - AAAA - A and AAAA - SRV - CNAME - Optionally, a health check After you create the service, you can submit a RegisterInstance request, and AWS Cloud Map uses the values in the configuration to create the specified entities\cr
\link[=servicediscovery_delete_namespace]{delete_namespace} \tab Deletes a namespace from the current account \cr
\link[=servicediscovery_delete_service]{delete_service} \tab Deletes a specified service \cr
\link[=servicediscovery_deregister_instance]{deregister_instance} \tab Deletes the Amazon Route 53 DNS records and health check, if any, that AWS Cloud Map created for the specified instance \cr
\link[=servicediscovery_discover_instances]{discover_instances} \tab Discovers registered instances for a specified namespace and service \cr
\link[=servicediscovery_get_instance]{get_instance} \tab Gets information about a specified instance \cr
\link[=servicediscovery_get_instances_health_status]{get_instances_health_status} \tab Gets the current health status (Healthy, Unhealthy, or Unknown) of one or more instances that are associated with a specified service \cr
\link[=servicediscovery_get_namespace]{get_namespace} \tab Gets information about a namespace \cr
\link[=servicediscovery_get_operation]{get_operation} \tab Gets information about any operation that returns an operation ID in the response, such as a CreateService request \cr
\link[=servicediscovery_get_service]{get_service} \tab Gets the settings for a specified service \cr
\link[=servicediscovery_list_instances]{list_instances} \tab Lists summary information about the instances that you registered by using a specified service \cr
\link[=servicediscovery_list_namespaces]{list_namespaces} \tab Lists summary information about the namespaces that were created by the current AWS account \cr
\link[=servicediscovery_list_operations]{list_operations} \tab Lists operations that match the criteria that you specify \cr
\link[=servicediscovery_list_services]{list_services} \tab Lists summary information for all the services that are associated with one or more specified namespaces \cr
\link[=servicediscovery_register_instance]{register_instance} \tab Creates or updates one or more records and, optionally, creates a health check based on the settings in a specified service \cr
\link[=servicediscovery_update_instance_custom_health_status]{update_instance_custom_health_status} \tab Submits a request to change the health status of a custom health check to healthy or unhealthy \cr
\link[=servicediscovery_update_service]{update_service} \tab Submits a request to perform the following operations: - Add or delete DnsRecords configurations - Update the TTL setting for existing DnsRecords configurations - Add, update, or delete HealthCheckConfig for a specified service For public and private DNS namespaces, you must specify all DnsRecords configurations (and, optionally, HealthCheckConfig) that you want to appear in the updated service
}
}

\examples{
\donttest{svc <- servicediscovery()
svc$create_http_namespace(
  Foo = 123
)}

}
