% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_aws_default_service_quota}
\alias{servicequotas_get_aws_default_service_quota}
\title{Retrieves the default value for the specified quota}
\usage{
servicequotas_get_aws_default_service_quota(ServiceCode, QuotaCode)
}
\arguments{
\item{ServiceCode}{[required] The service identifier.}

\item{QuotaCode}{[required] The quota identifier.}
}
\description{
Retrieves the default value for the specified quota. The default value
does not reflect any quota increases.
}
\section{Request syntax}{
\preformatted{svc$get_aws_default_service_quota(
  ServiceCode = "string",
  QuotaCode = "string"
)
}
}

\keyword{internal}
