% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_suppressed_destination}
\alias{sesv2_get_suppressed_destination}
\title{Used to fetch a single suppressed email destination from your
suppression list}
\usage{
sesv2_get_suppressed_destination(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] Email destination to fetch from the suppression list.}
}
\description{
Used to fetch a single suppressed email destination from your
suppression list.
}
\section{Request syntax}{
\preformatted{svc$get_suppressed_destination(
  EmailAddress = "string"
)
}
}

\keyword{internal}
