% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_suppressed_destination}
\alias{sesv2_put_suppressed_destination}
\title{Adds an email address to the suppression list for your account}
\usage{
sesv2_put_suppressed_destination(EmailAddress, Reason)
}
\arguments{
\item{EmailAddress}{[required] The email address that should be added to the suppression list for your
account.}

\item{Reason}{[required] The factors that should cause the email address to be added to the
suppression list for your account.}
}
\description{
Adds an email address to the suppression list for your account.
}
\section{Request syntax}{
\preformatted{svc$put_suppressed_destination(
  EmailAddress = "string",
  Reason = "BOUNCE"|"COMPLAINT"
)
}
}

\keyword{internal}
