% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_service.R
\name{signer}
\alias{signer}
\title{AWS Signer}
\usage{
signer(config = NULL)
}
\arguments{
\item{config}{An optional list of custom configurations for the service. Currently
supports adding custom credentials, endpoint, and region.}
}
\description{
You can use Code Signing for Amazon FreeRTOS (AWS Signer) to sign code
that you created for any of the IoT devices that Amazon Web Services
supports. AWS Signer is integrated with Amazon FreeRTOS, AWS Certificate
Manager, and AWS CloudTrail. Amazon FreeRTOS customers can use AWS
Signer to sign code images before making them available for
microcontrollers. You can use ACM to import third-party certificates to
be used by AWS Signer. For general information about using AWS Signer,
see the \href{http://docs.aws.amazon.com/signer/latest/developerguide/Welcome.html}{Code Signing for Amazon FreeRTOS Developer Guide}.
}
\section{Service syntax}{
\preformatted{svc <- signer(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string",
        provider_name = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=signer_cancel_signing_profile]{cancel_signing_profile} \tab Changes the state of an ACTIVE signing profile to CANCELED \cr
\link[=signer_describe_signing_job]{describe_signing_job} \tab Returns information about a specific code signing job \cr
\link[=signer_get_signing_platform]{get_signing_platform} \tab Returns information on a specific signing platform \cr
\link[=signer_get_signing_profile]{get_signing_profile} \tab Returns information on a specific signing profile \cr
\link[=signer_list_signing_jobs]{list_signing_jobs} \tab Lists all your signing jobs \cr
\link[=signer_list_signing_platforms]{list_signing_platforms} \tab Lists all signing platforms available in AWS Signer that match the request parameters\cr
\link[=signer_list_signing_profiles]{list_signing_profiles} \tab Lists all available signing profiles in your AWS account \cr
\link[=signer_put_signing_profile]{put_signing_profile} \tab Creates a signing profile \cr
\link[=signer_start_signing_job]{start_signing_job} \tab Initiates a signing job to be performed on the code provided
}
}

\examples{
\donttest{svc <- signer()
svc$cancel_signing_profile(
  Foo = 123
)}

}
