% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app}
\alias{sms_get_app}
\title{Retrieve information about an application}
\usage{
sms_get_app(appId)
}
\arguments{
\item{appId}{ID of the application whose information is being retrieved.}
}
\description{
Retrieve information about an application.
}
\section{Request syntax}{
\preformatted{svc$get_app(
  appId = "string"
)
}
}

\keyword{internal}
