% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_document}
\alias{ssm_delete_document}
\title{Deletes the Systems Manager document and all instance associations to
the document}
\usage{
ssm_delete_document(Name)
}
\arguments{
\item{Name}{[required] The name of the document.}
}
\description{
Deletes the Systems Manager document and all instance associations to
the document.
}
\details{
Before you delete the document, we recommend that you use
DeleteAssociation to disassociate all instances that are associated with
the document.
}
\section{Request syntax}{
\preformatted{ssm$delete_document(
  Name = "string"
)
}
}

\keyword{internal}
