% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_maintenance_window}
\alias{ssm_get_maintenance_window}
\title{Retrieves a Maintenance Window}
\usage{
ssm_get_maintenance_window(WindowId)
}
\arguments{
\item{WindowId}{[required] The ID of the desired Maintenance Window.}
}
\description{
Retrieves a Maintenance Window.
}
\section{Request syntax}{
\preformatted{ssm$get_maintenance_window(
  WindowId = "string"
)
}
}

\keyword{internal}
