% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_maintenance_window_task}
\alias{ssm_get_maintenance_window_task}
\title{Lists the tasks in a Maintenance Window}
\usage{
ssm_get_maintenance_window_task(WindowId, WindowTaskId)
}
\arguments{
\item{WindowId}{[required] The Maintenance Window ID that includes the task to retrieve.}

\item{WindowTaskId}{[required] The Maintenance Window task ID to retrieve.}
}
\description{
Lists the tasks in a Maintenance Window.
}
\section{Request syntax}{
\preformatted{ssm$get_maintenance_window_task(
  WindowId = "string",
  WindowTaskId = "string"
)
}
}

\keyword{internal}
