% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_severity_levels}
\alias{support_describe_severity_levels}
\title{Returns the list of severity levels that you can assign to an AWS
Support case}
\usage{
support_describe_severity_levels(language)
}
\arguments{
\item{language}{The ISO 639-1 code for the language in which AWS provides support. AWS
Support currently supports English (\"en\") and Japanese (\"ja\").
Language parameters must be passed explicitly for operations that take
them.}
}
\description{
Returns the list of severity levels that you can assign to an AWS
Support case. The severity level for a case is also a field in the
CaseDetails data type included in any CreateCase request.
}
\section{Request syntax}{
\preformatted{svc$describe_severity_levels(
  language = "string"
)
}
}

\keyword{internal}
