% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_service.R
\name{synthetics}
\alias{synthetics}
\title{Synthetics}
\usage{
synthetics(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon CloudWatch Synthetics

You can use Amazon CloudWatch Synthetics to continually monitor your
services. You can create and manage \emph{canaries}, which are modular,
lightweight scripts that monitor your endpoints and APIs from the
outside-in. You can set up your canaries to run 24 hours a day, once per
minute. The canaries help you check the availability and latency of your
web services and troubleshoot anomalies by investigating load time data,
screenshots of the UI, logs, and metrics. The canaries seamlessly
integrate with CloudWatch ServiceLens to help you trace the causes of
impacted nodes in your applications. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/ServiceLens.html}{Using ServiceLens to Monitor the Health of Your Applications}
in the \emph{Amazon CloudWatch User Guide}.

Before you create and manage canaries, be aware of the security
considerations. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/servicelens_canaries_security.html}{Security Considerations for Synthetics Canaries}.
}
\section{Service syntax}{
\preformatted{svc <- synthetics(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=synthetics_create_canary]{create_canary} \tab Creates a canary\cr
\link[=synthetics_delete_canary]{delete_canary} \tab Permanently deletes the specified canary\cr
\link[=synthetics_describe_canaries]{describe_canaries} \tab This operation returns a list of the canaries in your account, along with full details about each canary\cr
\link[=synthetics_describe_canaries_last_run]{describe_canaries_last_run} \tab Use this operation to see information from the most recent run of each canary that you have created\cr
\link[=synthetics_describe_runtime_versions]{describe_runtime_versions} \tab Returns a list of Synthetics canary runtime versions\cr
\link[=synthetics_get_canary]{get_canary} \tab Retrieves complete information about one canary\cr
\link[=synthetics_get_canary_runs]{get_canary_runs} \tab Retrieves a list of runs for a specified canary\cr
\link[=synthetics_list_tags_for_resource]{list_tags_for_resource} \tab Displays the tags associated with a canary\cr
\link[=synthetics_start_canary]{start_canary} \tab Use this operation to run a canary that has already been created\cr
\link[=synthetics_stop_canary]{stop_canary} \tab Stops the canary to prevent all future runs\cr
\link[=synthetics_tag_resource]{tag_resource} \tab Assigns one or more tags (key-value pairs) to the specified canary\cr
\link[=synthetics_untag_resource]{untag_resource} \tab Removes one or more tags from the specified canary\cr
\link[=synthetics_update_canary]{update_canary} \tab Use this operation to change the settings of a canary that has already been created
}
}

\examples{
\dontrun{
svc <- synthetics()
svc$create_canary(
  Foo = 123
)
}

}
