% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_service.R
\name{timestreamquery}
\alias{timestreamquery}
\title{Amazon Timestream Query}
\usage{
timestreamquery(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Timestream Query
}
\section{Service syntax}{
\preformatted{svc <- timestreamquery(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=timestreamquery_cancel_query]{cancel_query} \tab Cancels a query that has been issued\cr
\link[=timestreamquery_describe_endpoints]{describe_endpoints} \tab DescribeEndpoints returns a list of available endpoints to make Timestream API calls against\cr
\link[=timestreamquery_query]{query} \tab Query is a synchronous operation that enables you to execute a query
}
}

\examples{
\dontrun{
svc <- timestreamquery()
svc$cancel_query(
  Foo = 123
)
}

}
