% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_service.R
\name{transcribeservice}
\alias{transcribeservice}
\title{Amazon Transcribe Service}
\usage{
transcribeservice(config = NULL)
}
\arguments{
\item{config}{An optional list of custom configurations for the service. Currently
supports adding custom credentials, endpoint, and region.}
}
\description{
Operations and objects for transcribing speech to text.
}
\section{Service syntax}{
\preformatted{svc <- transcribeservice(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string",
        provider_name = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=transcribeservice_create_vocabulary]{create_vocabulary} \tab Creates a new custom vocabulary that you can use to change the way Amazon Transcribe handles transcription of an audio file \cr
\link[=transcribeservice_delete_transcription_job]{delete_transcription_job} \tab Deletes a previously submitted transcription job along with any other generated results such as the transcription, models, and so on\cr
\link[=transcribeservice_delete_vocabulary]{delete_vocabulary} \tab Deletes a vocabulary from Amazon Transcribe \cr
\link[=transcribeservice_get_transcription_job]{get_transcription_job} \tab Returns information about a transcription job \cr
\link[=transcribeservice_get_vocabulary]{get_vocabulary} \tab Gets information about a vocabulary \cr
\link[=transcribeservice_list_transcription_jobs]{list_transcription_jobs} \tab Lists transcription jobs with the specified status \cr
\link[=transcribeservice_list_vocabularies]{list_vocabularies} \tab Returns a list of vocabularies that match the specified criteria \cr
\link[=transcribeservice_start_transcription_job]{start_transcription_job} \tab Starts an asynchronous job to transcribe speech to text \cr
\link[=transcribeservice_update_vocabulary]{update_vocabulary} \tab Updates an existing vocabulary with new values
}
}

\examples{
\donttest{svc <- transcribeservice()
svc$create_vocabulary(
  Foo = 123
)}

}
