% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{transcribeservice}
\alias{transcribeservice}
\title{Amazon Transcribe Service}
\usage{
transcribeservice(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Operations and objects for transcribing speech to text.
}
\section{Service syntax}{
\preformatted{svc <- transcribeservice(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=transcribeservice_create_medical_vocabulary]{create_medical_vocabulary} \tab Creates a new custom vocabulary that you can use to change how Amazon Transcribe Medical transcribes your audio file \cr
\link[=transcribeservice_create_vocabulary]{create_vocabulary} \tab Creates a new custom vocabulary that you can use to change the way Amazon Transcribe handles transcription of an audio file \cr
\link[=transcribeservice_create_vocabulary_filter]{create_vocabulary_filter} \tab Creates a new vocabulary filter that you can use to filter words, such as profane words, from the output of a transcription job \cr
\link[=transcribeservice_delete_medical_transcription_job]{delete_medical_transcription_job} \tab Deletes a transcription job generated by Amazon Transcribe Medical and any related information \cr
\link[=transcribeservice_delete_medical_vocabulary]{delete_medical_vocabulary} \tab Deletes a vocabulary from Amazon Transcribe Medical \cr
\link[=transcribeservice_delete_transcription_job]{delete_transcription_job} \tab Deletes a previously submitted transcription job along with any other generated results such as the transcription, models, and so on\cr
\link[=transcribeservice_delete_vocabulary]{delete_vocabulary} \tab Deletes a vocabulary from Amazon Transcribe \cr
\link[=transcribeservice_delete_vocabulary_filter]{delete_vocabulary_filter} \tab Removes a vocabulary filter \cr
\link[=transcribeservice_get_medical_transcription_job]{get_medical_transcription_job} \tab Returns information about a transcription job from Amazon Transcribe Medical \cr
\link[=transcribeservice_get_medical_vocabulary]{get_medical_vocabulary} \tab Retrieve information about a medical vocabulary \cr
\link[=transcribeservice_get_transcription_job]{get_transcription_job} \tab Returns information about a transcription job \cr
\link[=transcribeservice_get_vocabulary]{get_vocabulary} \tab Gets information about a vocabulary \cr
\link[=transcribeservice_get_vocabulary_filter]{get_vocabulary_filter} \tab Returns information about a vocabulary filter \cr
\link[=transcribeservice_list_medical_transcription_jobs]{list_medical_transcription_jobs} \tab Lists medical transcription jobs with a specified status or substring that matches their names \cr
\link[=transcribeservice_list_medical_vocabularies]{list_medical_vocabularies} \tab Returns a list of vocabularies that match the specified criteria \cr
\link[=transcribeservice_list_transcription_jobs]{list_transcription_jobs} \tab Lists transcription jobs with the specified status \cr
\link[=transcribeservice_list_vocabularies]{list_vocabularies} \tab Returns a list of vocabularies that match the specified criteria \cr
\link[=transcribeservice_list_vocabulary_filters]{list_vocabulary_filters} \tab Gets information about vocabulary filters \cr
\link[=transcribeservice_start_medical_transcription_job]{start_medical_transcription_job} \tab Start a batch job to transcribe medical speech to text \cr
\link[=transcribeservice_start_transcription_job]{start_transcription_job} \tab Starts an asynchronous job to transcribe speech to text \cr
\link[=transcribeservice_update_medical_vocabulary]{update_medical_vocabulary} \tab Updates an existing vocabulary with new values in a different text file \cr
\link[=transcribeservice_update_vocabulary]{update_vocabulary} \tab Updates an existing vocabulary with new values \cr
\link[=transcribeservice_update_vocabulary_filter]{update_vocabulary_filter} \tab Updates a vocabulary filter with a new list of filtered words
}
}

\examples{
\dontrun{
svc <- transcribeservice()
svc$create_medical_vocabulary(
  Foo = 123
)
}

}
