% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_list_tags_for_resource}
\alias{transfer_list_tags_for_resource}
\title{Lists all of the tags associated with the Amazon Resource Number (ARN)
you specify}
\usage{
transfer_list_tags_for_resource(Arn, MaxResults, NextToken)
}
\arguments{
\item{Arn}{[required] Requests the tags associated with a particular Amazon Resource Name
(ARN). An ARN is an identifier for a specific AWS resource, such as a
server, user, or role.}

\item{MaxResults}{Specifies the number of tags to return as a response to the
\code{ListTagsForResource} request.}

\item{NextToken}{When you request additional results from the \code{ListTagsForResource} call,
a \code{NextToken} parameter is returned in the input. You can then pass in a
subsequent command the \code{NextToken} parameter to continue listing
additional tags.}
}
\description{
Lists all of the tags associated with the Amazon Resource Number (ARN)
you specify. The resource can be a user, server, or role.
}
\section{Request syntax}{
\preformatted{transfer$list_tags_for_resource(
  Arn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
