% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.R
\name{translate}
\alias{translate}
\title{Amazon Translate}
\usage{
translate()
}
\description{
Provides translation between one source language and another of the same
set of languages.
}
\section{Operations}{

\tabular{ll}{
\link[=translate_delete_terminology]{delete_terminology} \tab A synchronous action that deletes a custom terminology \cr
\link[=translate_get_terminology]{get_terminology} \tab Retrieves a custom terminology \cr
\link[=translate_import_terminology]{import_terminology} \tab Creates or updates a custom terminology, depending on whether or not one already exists for the given terminology name\cr
\link[=translate_list_terminologies]{list_terminologies} \tab Provides a list of custom terminologies associated with your account \cr
\link[=translate_translate_text]{translate_text} \tab Translates input text from the source language to the target language
}
}

\examples{
\donttest{svc <- translate()
svc$delete_terminology(
  Foo = 123
)}

}
