% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_logging_configuration}
\alias{waf_get_logging_configuration}
\title{This is AWS WAF Classic documentation}
\usage{
waf_get_logging_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the web ACL for which you want to get
the LoggingConfiguration.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Returns the LoggingConfiguration for the specified web ACL.
}
\section{Request syntax}{
\preformatted{svc$get_logging_configuration(
  ResourceArn = "string"
)
}
}

\keyword{internal}
