% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_sql_injection_match_set}
\alias{wafregional_get_sql_injection_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_get_sql_injection_match_set(SqlInjectionMatchSetId)
}
\arguments{
\item{SqlInjectionMatchSetId}{[required] The \code{SqlInjectionMatchSetId} of the SqlInjectionMatchSet that you want
to get. \code{SqlInjectionMatchSetId} is returned by
CreateSqlInjectionMatchSet and by ListSqlInjectionMatchSets.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Returns the SqlInjectionMatchSet that is specified by
\code{SqlInjectionMatchSetId}.
}
\section{Request syntax}{
\preformatted{svc$get_sql_injection_match_set(
  SqlInjectionMatchSetId = "string"
)
}
}

\examples{
\dontrun{
# The following example returns the details of a SQL injection match set
# with the ID example1ds3t-46da-4fdb-b8d5-abc321j569j5.
svc$get_sql_injection_match_set(
  SqlInjectionMatchSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)
}

}
\keyword{internal}
