% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_service.R
\name{wellarchitected}
\alias{wellarchitected}
\title{AWS Well-Architected Tool}
\usage{
wellarchitected(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
This is the \emph{AWS Well-Architected Tool API Reference}.

The AWS Well-Architected Tool API provides programmatic access to the
\href{https://aws.amazon.com/well-architected-tool/}{AWS Well-Architected Tool} in the \href{https://console.aws.amazon.com/wellarchitected/home}{AWS Management Console}.

\strong{Managing workloads:}
\itemize{
\item CreateWorkload: Define a new workload.
\item ListWorkloads: List workloads.
\item GetWorkload: Get the properties of a workload.
\item UpdateWorkload: Update the properties of a workload.
\item DeleteWorkload: Delete a workload.
}

\strong{Managing milestones:}
\itemize{
\item CreateMilestone: Create a milestone.
\item ListMilestones: List milestones.
\item GetMilestone: Get the properties of a milestone.
}

\strong{Managing lenses:}
\itemize{
\item ListLenses: List the available lenses.
\item AssociateLenses: Add one or more lenses to a workload.
\item DisassociateLenses: Remove one or more lenses from a workload.
\item ListNotifications: List lens notifications for a workload.
\item GetLensVersionDifference: Get the differences between the version of
a lens used in a workload and the latest version.
\item UpgradeLensReview: Upgrade a workload to use the latest version of a
lens.
}

\strong{Managing reviews:}
\itemize{
\item ListLensReviews: List reviews associated with a workload.
\item GetLensReview: Get pillar and risk data associated with a workload
review.
\item GetLensReviewReport: Get the report associated with a workload
review.
\item UpdateLensReview: Update the notes associated with a workload
review.
\item ListAnswers: List the questions and answers in a workload review.
\item UpdateAnswer: Update the answer to a specific question in a workload
review.
\item ListLensReviewImprovements: List the improvement plan associated
with a workload review.
}

\strong{Managing workload shares:}
\itemize{
\item ListWorkloadShares: List the workload shares associated with a
workload.
\item CreateWorkloadShare: Create a workload share.
\item UpdateWorkloadShare: Update a workload share.
\item DeleteWorkloadShare: Delete a workload share.
}

\strong{Managing workload share invitations:}
\itemize{
\item ListShareInvitations: List workload share invitations.
\item UpdateShareInvitation: Update a workload share invitation.
}

For information about the AWS Well-Architected Tool, see the \href{https://docs.aws.amazon.com/wellarchitected/latest/userguide/intro.html}{AWS Well-Architected Tool User Guide}.
}
\section{Service syntax}{
\preformatted{svc <- wellarchitected(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=wellarchitected_associate_lenses]{associate_lenses} \tab Associate a lens to a workload\cr
\link[=wellarchitected_create_milestone]{create_milestone} \tab Create a milestone for an existing workload\cr
\link[=wellarchitected_create_workload]{create_workload} \tab Create a new workload\cr
\link[=wellarchitected_create_workload_share]{create_workload_share} \tab Create a workload share\cr
\link[=wellarchitected_delete_workload]{delete_workload} \tab Delete an existing workload\cr
\link[=wellarchitected_delete_workload_share]{delete_workload_share} \tab Delete a workload share\cr
\link[=wellarchitected_disassociate_lenses]{disassociate_lenses} \tab Disassociate a lens from a workload\cr
\link[=wellarchitected_get_answer]{get_answer} \tab Get lens review\cr
\link[=wellarchitected_get_lens_review]{get_lens_review} \tab Get lens review\cr
\link[=wellarchitected_get_lens_review_report]{get_lens_review_report} \tab Get lens review report\cr
\link[=wellarchitected_get_lens_version_difference]{get_lens_version_difference} \tab Get lens version differences\cr
\link[=wellarchitected_get_milestone]{get_milestone} \tab Get a milestone for an existing workload\cr
\link[=wellarchitected_get_workload]{get_workload} \tab Get an existing workload\cr
\link[=wellarchitected_list_answers]{list_answers} \tab List of answers\cr
\link[=wellarchitected_list_lenses]{list_lenses} \tab List the available lenses\cr
\link[=wellarchitected_list_lens_review_improvements]{list_lens_review_improvements} \tab List lens review improvements\cr
\link[=wellarchitected_list_lens_reviews]{list_lens_reviews} \tab List lens reviews\cr
\link[=wellarchitected_list_milestones]{list_milestones} \tab List all milestones for an existing workload\cr
\link[=wellarchitected_list_notifications]{list_notifications} \tab List lens notifications\cr
\link[=wellarchitected_list_share_invitations]{list_share_invitations} \tab List the workload invitations\cr
\link[=wellarchitected_list_workloads]{list_workloads} \tab List workloads\cr
\link[=wellarchitected_list_workload_shares]{list_workload_shares} \tab List the workload shares associated with the workload\cr
\link[=wellarchitected_update_answer]{update_answer} \tab Update the answer\cr
\link[=wellarchitected_update_lens_review]{update_lens_review} \tab Update lens review\cr
\link[=wellarchitected_update_share_invitation]{update_share_invitation} \tab Update a workload invitation\cr
\link[=wellarchitected_update_workload]{update_workload} \tab Update an existing workload\cr
\link[=wellarchitected_update_workload_share]{update_workload_share} \tab Update a workload share\cr
\link[=wellarchitected_upgrade_lens_review]{upgrade_lens_review} \tab Upgrade lens review
}
}

\examples{
\dontrun{
svc <- wellarchitected()
svc$associate_lenses(
  Foo = 123
)
}

}
