% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmailmessageflow_service.R
\name{workmailmessageflow}
\alias{workmailmessageflow}
\title{Amazon WorkMail Message Flow}
\usage{
workmailmessageflow(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
The WorkMail Message Flow API provides access to email messages as they
are being sent and received by a WorkMail organization.
}
\section{Service syntax}{
\preformatted{svc <- workmailmessageflow(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Examples}{
\preformatted{svc$get_raw_message_content(
  Foo = 123
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=workmailmessageflow_get_raw_message_content]{get_raw_message_content} \tab Retrieves the raw content of an in-transit email message, in MIME format
}
}

