% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_service.R
\name{workspaces}
\alias{workspaces}
\title{Amazon WorkSpaces}
\usage{
workspaces()
}
\description{
Amazon WorkSpaces Service

Amazon WorkSpaces enables you to provision virtual, cloud-based
Microsoft Windows and Amazon Linux desktops for your users.
}
\section{Example}{
\preformatted{workspaces <- paws::workspaces()
workspaces$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=workspaces_associate_ip_groups]{associate_ip_groups} \tab Associates the specified IP access control group with the specified directory \cr
\link[=workspaces_authorize_ip_rules]{authorize_ip_rules} \tab Adds one or more rules to the specified IP access control group \cr
\link[=workspaces_create_ip_group]{create_ip_group} \tab Creates an IP access control group \cr
\link[=workspaces_create_tags]{create_tags} \tab Creates the specified tags for the specified WorkSpace \cr
\link[=workspaces_create_workspaces]{create_workspaces} \tab Creates one or more WorkSpaces \cr
\link[=workspaces_delete_ip_group]{delete_ip_group} \tab Deletes the specified IP access control group \cr
\link[=workspaces_delete_tags]{delete_tags} \tab Deletes the specified tags from the specified WorkSpace \cr
\link[=workspaces_delete_workspace_image]{delete_workspace_image} \tab Deletes the specified image from your account \cr
\link[=workspaces_describe_account]{describe_account} \tab Retrieves a list that describes the configuration of bring your own license (BYOL) for the specified account \cr
\link[=workspaces_describe_account_modifications]{describe_account_modifications} \tab Retrieves a list that describes modifications to the configuration of bring your own license (BYOL) for the specified account \cr
\link[=workspaces_describe_client_properties]{describe_client_properties} \tab Retrieves a list that describes one or more specified Amazon WorkSpaces clients \cr
\link[=workspaces_describe_ip_groups]{describe_ip_groups} \tab Describes one or more of your IP access control groups \cr
\link[=workspaces_describe_tags]{describe_tags} \tab Describes the specified tags for the specified WorkSpace \cr
\link[=workspaces_describe_workspace_bundles]{describe_workspace_bundles} \tab Retrieves a list that describes the available WorkSpace bundles \cr
\link[=workspaces_describe_workspace_directories]{describe_workspace_directories} \tab Describes the available AWS Directory Service directories that are registered with Amazon WorkSpaces \cr
\link[=workspaces_describe_workspace_images]{describe_workspace_images} \tab Retrieves a list that describes one or more specified images, if the image identifiers are provided \cr
\link[=workspaces_describe_workspaces]{describe_workspaces} \tab Describes the specified WorkSpaces \cr
\link[=workspaces_describe_workspaces_connection_status]{describe_workspaces_connection_status} \tab Describes the connection status of the specified WorkSpaces \cr
\link[=workspaces_disassociate_ip_groups]{disassociate_ip_groups} \tab Disassociates the specified IP access control group from the specified directory \cr
\link[=workspaces_import_workspace_image]{import_workspace_image} \tab Imports the specified Windows 7 or Windows 10 bring your own license (BYOL) image into Amazon WorkSpaces \cr
\link[=workspaces_list_available_management_cidr_ranges]{list_available_management_cidr_ranges} \tab Retrieves a list of IP address ranges, specified as IPv4 CIDR blocks, that you can use for the network management interface when you enable bring your own license (BYOL)\cr
\link[=workspaces_modify_account]{modify_account} \tab Modifies the configuration of bring your own license (BYOL) for the specified account \cr
\link[=workspaces_modify_client_properties]{modify_client_properties} \tab Modifies the properties of the specified Amazon WorkSpaces client \cr
\link[=workspaces_modify_workspace_properties]{modify_workspace_properties} \tab Modifies the specified WorkSpace properties \cr
\link[=workspaces_modify_workspace_state]{modify_workspace_state} \tab Sets the state of the specified WorkSpace \cr
\link[=workspaces_reboot_workspaces]{reboot_workspaces} \tab Reboots the specified WorkSpaces \cr
\link[=workspaces_rebuild_workspaces]{rebuild_workspaces} \tab Rebuilds the specified WorkSpace \cr
\link[=workspaces_revoke_ip_rules]{revoke_ip_rules} \tab Removes one or more rules from the specified IP access control group \cr
\link[=workspaces_start_workspaces]{start_workspaces} \tab Starts the specified WorkSpaces \cr
\link[=workspaces_stop_workspaces]{stop_workspaces} \tab Stops the specified WorkSpaces \cr
\link[=workspaces_terminate_workspaces]{terminate_workspaces} \tab Terminates the specified WorkSpaces \cr
\link[=workspaces_update_rules_of_ip_group]{update_rules_of_ip_group} \tab Replaces the current rules of the specified IP access control group with the specified rules
}
}

